Get the upstream and downstream connectivity of a neuron

Get the upstream and downstream connectivity of a body, restricted to within an ROI if specified

neuprint_connection_table(
  bodyids,
  prepost = c("PRE", "POST"),
  roi = NULL,
  progress = FALSE,
  dataset = NULL,
  all_segments = FALSE,
  conn = NULL,
  ...
)

Arguments

bodyids
the body IDs for neurons/segments (bodies) you wish to query

prepost
whether to look for partners presynaptic to postsynaptic to the given bodyids

roi
a single ROI. Use neuprint_ROIs to see what is available.

progress
default FALSE. If TRUE, the API is called separately for each neuron and you can assess its progress, if an error is thrown by any one bodyid, that bodyid is ignored

dataset
optional, a dataset you want to query. If NULL, the default specified by your R environ file is used. See neuprint_login for details.

all_segments
if TRUE, all bodies are considered, if FALSE, only 'Neurons', i.e. bodies with a status roughly traced status.

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

Value

a data frame giving partners within a set of ROIs, the connection strength for weights to or from that partner, and the direction, for the given bodyid

See also

Examples

neuprint_connection_table(c(818983130, 1796818119))

#>         bodyid    partner         roi weight prepost
#> 1    818983130 5901202409       AL(R)     28       0
#> 220 1796818119 5901202409       AL(R)     28       0
#> 301 1796818119 1825789179       AL(R)     25       0
#> 95   818983130 1825789179       AL(R)     24       0
#> 337 1796818119 1671620613       AL(R)     22       0
#> 309 1796818119 1796813992       AL(R)     17       0
#> 332 1796818119 1700393451       AL(R)     17       0
#> 108  818983130 1796809680       AL(R)     16       0
#> 282 1796818119 1859224229       AL(R)     16       0
#> 326 1796818119 1702802179       AL(R)     16       0
#> 267 1796818119 1858171556       AL(R)     14       0
#> 338 1796818119 1670390647       AL(R)     14       0
#> 4    818983130 5901198995       AL(R)     12       0
#> 131  818983130 1671620613       AL(R)     11       0
#> 311 1796818119 1796809680       AL(R)     11       0
#> 330 1796818119 1702306037       AL(R)     11       0
#> 61   818983130 1889931525       AL(R)     10       0
#> 194  818983130  573695941       LH(R)     10       0
#> 240 1796818119 5813040847       AL(R)     10       0
#> 265 1796818119 1858892359       AL(R)     10       0
#> 279 1796818119 1859224739       AL(R)     10       0
#> 307 1796818119 1796818061       AL(R)     10       0
#> 325 1796818119 1733833296       AL(R)     10       0
#> 266 1796818119 1858887650       AL(R)      9       0
#> 284 1796818119 1859220700       AL(R)      9       0
#> 290 1796818119 1828189572       AL(R)      9       0
#> 124  818983130 1702306037       AL(R)      8       0
#> 150  818983130 1670627928       AL(R)      8       0
#> 237 1796818119 5813054726       AL(R)      8       0
#> 263 1796818119 1889931525       AL(R)      8       0
#> 335 1796818119 1701675592       AL(R)      8       0
#> 339 1796818119 1670386270       AL(R)      8       0
#> 352 1796818119 1670627928       AL(R)      8       0
#> 66   818983130 1858892359       AL(R)      7       0
#> 133  818983130 1670386270       AL(R)      7       0
#> 232 1796818119 5813055277       AL(R)      7       0
#> 236 1796818119 5813054773       AL(R)      7       0
#> 280 1796818119 1859224592       AL(R)      7       0
#> 289 1796818119 1828194121       AL(R)      7       0
#> 306 1796818119 1796818191       AL(R)      7       0
#> 310 1796818119 1796813596       AL(R)      7       0
#> 416 1796818119  513323340       LH(R)      7       0
#> 74   818983130 1859565686       AL(R)      6       0
#> 101  818983130 1796818191       AL(R)      6       0
#> 120  818983130 1702802179       AL(R)      6       0
#> 125  818983130 1700393451       AL(R)      6       0
#> 144  818983130 1671659809       AL(R)      6       0
#> 206  818983130  487144598       LH(R)      6       0
#> 274 1796818119 1859565686       AL(R)      6       0
#> 283 1796818119 1859220826       AL(R)      6       0
#> 304 1796818119 1797505019       AL(R)      6       0
#> 320 1796818119 1732995501       AL(R)      6       0
#> 355 1796818119 1667251683       AL(R)      6       0
#> 359 1796818119 1640922754       AL(R)      6       0
#> 432 1796818119  425790257 MB(+ACA)(R)      6       0
#> 433 1796818119  425790257       MB(R)      6       0
#> 436 1796818119  425790257       CA(R)      6       0
#> 86   818983130 1828194121       AL(R)      5       0
#> 94   818983130 1826445251       AL(R)      5       0
#> 96   818983130 1797505019       LH(R)      5       0
#> 105  818983130 1796817841       LH(R)      5       0
#> 119  818983130 1733833296       AL(R)      5       0
#> 145  818983130 1669946117       AL(R)      5       0
#> 155  818983130 1639886198       AL(R)      5       0
#> 162  818983130 1610530558       AL(R)      5       0
#> 201  818983130  513323340       LH(R)      5       0
#> 214  818983130  425790257 MB(+ACA)(R)      5       0
#> 215  818983130  425790257       MB(R)      5       0
#> 216  818983130  425790257       CA(R)      5       0
#> 227 1796818119 5813078440       AL(R)      5       0
#> 242 1796818119 5813032595       AL(R)      5       0
#> 256 1796818119 1920258429       AL(R)      5       0
#> 278 1796818119 1859224808       AL(R)      5       0
#> 293 1796818119 1827852654       AL(R)      5       0
#> 298 1796818119 1826929327       AL(R)      5       0
#> 303 1796818119 1797505019       LH(R)      5       0
#> 313 1796818119 1795799176       AL(R)      5       0
#> 323 1796818119 1730327790       AL(R)      5       0
#> 360 1796818119 1640922516       AL(R)      5       0
#> 409 1796818119  573695941       LH(R)      5       0
#> 25   818983130 5813054726       AL(R)      4       0
#> 52   818983130 1920957731       AL(R)      4       0
#> 68   818983130 1858171556       AL(R)      4       0
#> 81   818983130 1859220826       AL(R)      4       0
#> 102  818983130 1796818119       LH(R)      4       0
#> 113  818983130 1733491818       AL(R)      4       0
#> 117  818983130 1730327790       AL(R)      4       0
#> 149  818983130 1669699714       AL(R)      4       0
#> 153  818983130 1667251683       AL(R)      4       0
#> 202  818983130  512982243       LH(R)      4       0
#> 218 1796818119 5901218894       AL(R)      4       0
#> 222 1796818119 5901198995       AL(R)      4       0
#> 229 1796818119 5813069055       AL(R)      4       0
#> 246 1796818119 5812994702       AL(R)      4       0
#> 251 1796818119 2105086391       AL(R)      4       0
#> 257 1796818119 1919208947       AL(R)      4       0
#> 296 1796818119 1827516355       LH(R)      4       0
#> 305 1796818119 1796822470       AL(R)      4       0
#> 333 1796818119 1702107389       AL(R)      4       0
#> 340 1796818119 1670382036       AL(R)      4       0
#> 341 1796818119 1672320134       AL(R)      4       0
#> 342 1796818119 1671257931       AL(R)      4       0
#> 357 1796818119 1639886198       AL(R)      4       0
#> 365 1796818119 1610530558       AL(R)      4       0
#> 369 1796818119  947453617       LH(R)      4       0
#> 384 1796818119  818983130       LH(R)      4       0
#> 10   818983130 5813055277       AL(R)      3       0
#> 27   818983130 5813040847       AL(R)      3       0
#> 32   818983130 5813032595       AL(R)      3       0
#> 64   818983130 1889927339       AL(R)      3       0
#> 67   818983130 1858887650       AL(R)      3       0
#> 77   818983130 1859224808       AL(R)      3       0
#> 88   818983130 1828189572       AL(R)      3       0
#> 90   818983130 1824101645       AL(R)      3       0
#> 93   818983130 1826929327       AL(R)      3       0
#> 99   818983130 1796822470       AL(R)      3       0
#> 103  818983130 1796818119       AL(R)      3       0
#> 127  818983130 1701947337       AL(R)      3       0
#> 128  818983130 1701757711       AL(R)      3       0
#> 129  818983130 1701675592       AL(R)      3       0
#> 132  818983130 1670390647       AL(R)      3       0
#> 152  818983130 1669017843       AL(R)      3       0
#> 154  818983130 1640555311       AL(R)      3       0
#> 185  818983130  698194389       LH(R)      3       0
#> 198  818983130  544021095       LH(R)      3       0
#> 223 1796818119 5813130064       AL(R)      3       0
#> 230 1796818119 5813062199       AL(R)      3       0
#> 235 1796818119 5813054777       AL(R)      3       0
#> 238 1796818119 5813039491       AL(R)      3       0
#> 241 1796818119 5813039309       AL(R)      3       0
#> 260 1796818119 1888524453       LH(R)      3       0
#> 270 1796818119 1859574636       AL(R)      3       0
#> 275 1796818119 1859565147       AL(R)      3       0
#> 286 1796818119 1828530446       AL(R)      3       0
#> 288 1796818119 1828198909       AL(R)      3       0
#> 299 1796818119 1826445251       AL(R)      3       0
#> 317 1796818119 1762354941       AL(R)      3       0
#> 318 1796818119 1733491818       AL(R)      3       0
#> 321 1796818119 1732797061       AL(R)      3       0
#> 327 1796818119 1702651358       AL(R)      3       0
#> 353 1796818119 1669358817       AL(R)      3       0
#> 354 1796818119 1669017843       AL(R)      3       0
#> 356 1796818119 1640555311       AL(R)      3       0
#> 358 1796818119 1642623277       AL(R)      3       0
#> 361 1796818119 1640909284       AL(R)      3       0
#> 366 1796818119 1547454812       AL(R)      3       0
#> 368 1796818119  946814479       LH(R)      3       0
#> 377 1796818119  851317951       LH(R)      3       0
#> 2    818983130 5901203780       LH(R)      2       0
#> 3    818983130 5813130064       AL(R)      2       0
#> 9    818983130 5813062199       AL(R)      2       0
#> 19   818983130 5813054911       AL(R)      2       0
#> 20   818983130 5813054725       AL(R)      2       0
#> 24   818983130 5813054773       AL(R)      2       0
#> 29   818983130 5813039309       AL(R)      2       0
#> 43   818983130 2041621893       AL(R)      2       0
#> 45   818983130 1921307609       AL(R)      2       0
#> 46   818983130 1921307415       AL(R)      2       0
#> 49   818983130 1921303555       AL(R)      2       0
#> 50   818983130 1921302674       AL(R)      2       0
#> 51   818983130 1921298843       AL(R)      2       0
#> 57   818983130 1888524453       LH(R)      2       0
#> 60   818983130 1890272574       AL(R)      2       0
#> 62   818983130 1889931382       AL(R)      2       0
#> 78   818983130 1859224739       AL(R)      2       0
#> 80   818983130 1828530446       AL(R)      2       0
#> 85   818983130 1828198384       AL(R)      2       0
#> 89   818983130 1827852654       AL(R)      2       0
#> 92   818983130 1827516355       LH(R)      2       0
#> 100  818983130 1796818363       AL(R)      2       0
#> 104  818983130 1796818061       AL(R)      2       0
#> 106  818983130 1796817841       AL(R)      2       0
#> 107  818983130 1796813992       AL(R)      2       0
#> 110  818983130 1762704980       AL(R)      2       0
#> 111  818983130 1762359683       AL(R)      2       0
#> 115  818983130 1732995501       AL(R)      2       0
#> 116  818983130 1732792926       AL(R)      2       0
#> 126  818983130 1702107389       AL(R)      2       0
#> 134  818983130 1670287030       AL(R)      2       0
#> 136  818983130 1671270875       AL(R)      2       0
#> 137  818983130 1672320134       AL(R)      2       0
#> 148  818983130 1670916819       AL(R)      2       0
#> 159  818983130 1640909284       AL(R)      2       0
#> 177  818983130  851317951       LH(R)      2       0
#> 181  818983130  761938645       LH(R)      2       0
#> 183  818983130  724816115        None      2       0
#> 184  818983130  724816115     mALT(R)      2       0
#> 190  818983130  637121560       LH(R)      2       0
#> 211  818983130  480590566       LH(R)      2       0
#> 225 1796818119 5813086006       AL(R)      2       0
#> 231 1796818119 5813056598       AL(R)      2       0
#> 234 1796818119 5813054622       AL(R)      2       0
#> 253 1796818119 1921307609       AL(R)      2       0
#> 254 1796818119 1917188956        None      2       0
#> 255 1796818119 1917188956     mALT(R)      2       0
#> 269 1796818119 1859915060       AL(R)      2       0
#> 276 1796818119 1859229253       AL(R)      2       0
#> 291 1796818119 1828185706       AL(R)      2       0
#> 295 1796818119 1824101645       AL(R)      2       0
#> 316 1796818119 1762359683       AL(R)      2       0
#> 322 1796818119 1732792926       AL(R)      2       0
#> 324 1796818119 1704347707       AL(R)      2       0
#> 350 1796818119 1670916819       AL(R)      2       0
#> 370 1796818119  944729312       LH(R)      2       0
#> 373 1796818119  916763916       LH(R)      2       0
#> 380 1796818119  852678239       LH(R)      2       0
#> 383 1796818119  819220237       LH(R)      2       0
#> 390 1796818119  759810119       AL(R)      2       0
#> 392 1796818119  759810119     mALT(R)      2       0
#> 393 1796818119  759810119        None      2       0
#> 395 1796818119  726207450 MB(+ACA)(R)      2       0
#> 396 1796818119  726207450       MB(R)      2       0
#> 397 1796818119  726207450       CA(R)      2       0
#> 400 1796818119  698194389       LH(R)      2       0
#> 403 1796818119  637121560       LH(R)      2       0
#> 404 1796818119  635416407        None      2       0
#> 405 1796818119  635416407     mALT(R)      2       0
#> 412 1796818119  543679473       LH(R)      2       0
#> 421 1796818119  487144598       LH(R)      2       0
#> 425 1796818119  478613576 MB(+ACA)(R)      2       0
#> 426 1796818119  478613576       MB(R)      2       0
#> 428 1796818119  478613576       LH(R)      2       0
#> 430 1796818119  478613576       CA(R)      2       0
#> 5    818983130 5813076969       AL(R)      1       0
#> 6    818983130 5813078440       AL(R)      1       0
#> 7    818983130 5813069085       AL(R)      1       0
#> 8    818983130 5813069055       AL(R)      1       0
#> 11   818983130 5813056596        None      1       0
#> 12   818983130 5813056596     mALT(R)      1       0
#> 13   818983130 5813057869 MB(+ACA)(R)      1       0
#> 14   818983130 5813057869       MB(R)      1       0
#> 15   818983130 5813057869       CA(R)      1       0
#> 16   818983130 5813055180 MB(+ACA)(R)      1       0
#> 17   818983130 5813055180       MB(R)      1       0
#> 18   818983130 5813055180       CA(R)      1       0
#> 21   818983130 5813054622       AL(R)      1       0
#> 22   818983130 5813054777       AL(R)      1       0
#> 23   818983130 5813050455       LH(R)      1       0
#> 26   818983130 5813039491       AL(R)      1       0
#> 28   818983130 5813034455       AL(R)      1       0
#> 30   818983130 5813034306       AL(R)      1       0
#> 31   818983130 5813047691       AL(R)      1       0
#> 33   818983130 5813012083       LH(R)      1       0
#> 34   818983130 5813013443       LH(R)      1       0
#> 35   818983130 5812994702       AL(R)      1       0
#> 36   818983130 5812983069       AL(R)      1       0
#> 37   818983130 5812980220 MB(+ACA)(R)      1       0
#> 38   818983130 5812980220       MB(R)      1       0
#> 39   818983130 5812980220       CA(R)      1       0
#> 40   818983130 2069652905       AL(R)      1       0
#> 41   818983130 2071677193       AL(R)      1       0
#> 42   818983130 2040301572       AL(R)      1       0
#> 44   818983130 1977907527        None      1       0
#> 47   818983130 1921307399       AL(R)      1       0
#> 48   818983130 1921303672       AL(R)      1       0
#> 53   818983130 1920616720       AL(R)      1       0
#> 54   818983130 1949592261       AL(R)      1       0
#> 55   818983130 1920258429       AL(R)      1       0
#> 56   818983130 1920042851       AL(R)      1       0
#> 58   818983130 1890950476       AL(R)      1       0
#> 59   818983130 1890272912       AL(R)      1       0
#> 63   818983130 1889927761       AL(R)      1       0
#> 65   818983130 1887518225       AL(R)      1       0
#> 69   818983130 1856466057       AL(R)      1       0
#> 70   818983130 1851816763       AL(R)      1       0
#> 71   818983130 1859915060       AL(R)      1       0
#> 72   818983130 1859579279       AL(R)      1       0
#> 73   818983130 1859574636       AL(R)      1       0
#> 75   818983130 1859233517       AL(R)      1       0
#> 76   818983130 1859229253       AL(R)      1       0
#> 79   818983130 1859224229       AL(R)      1       0
#> 82   818983130 1828206773       AL(R)      1       0
#> 83   818983130 1828203598       AL(R)      1       0
#> 84   818983130 1828198909       AL(R)      1       0
#> 87   818983130 1828193782       AL(R)      1       0
#> 91   818983130 1827516355       AL(R)      1       0
#> 97   818983130 1797505019       AL(R)      1       0
#> 98   818983130 1797159438       AL(R)      1       0
#> 109  818983130 1795799176       AL(R)      1       0
#> 112  818983130 1762354941       AL(R)      1       0
#> 114  818983130 1733056086       AL(R)      1       0
#> 118  818983130 1704347707       AL(R)      1       0
#> 121  818983130 1702668580       AL(R)      1       0
#> 122  818983130 1702336197       AL(R)      1       0
#> 123  818983130 1702310266       AL(R)      1       0
#> 130  818983130 1699974843       AL(R)      1       0
#> 135  818983130 1671292719       AL(R)      1       0
#> 138  818983130 1671257931       AL(R)      1       0
#> 139  818983130 1670265095       AL(R)      1       0
#> 140  818983130 1670265016 MB(+ACA)(R)      1       0
#> 141  818983130 1670265016       MB(R)      1       0
#> 142  818983130 1670265016       AL(R)      1       0
#> 143  818983130 1670265016       CA(R)      1       0
#> 146  818983130 1670934213       LH(R)      1       0
#> 147  818983130 1667933908       AL(R)      1       0
#> 151  818983130 1669358817       AL(R)      1       0
#> 156  818983130 1642623277       AL(R)      1       0
#> 157  818983130 1640922754       AL(R)      1       0
#> 158  818983130 1640922516        None      1       0
#> 160  818983130 1636225502       AL(R)      1       0
#> 161  818983130 1640594274       AL(R)      1       0
#> 163  818983130 1547454812       AL(R)      1       0
#> 164  818983130 1447205712       LH(R)      1       0
#> 165  818983130 1297202769 MB(+ACA)(R)      1       0
#> 166  818983130 1297202769       MB(R)      1       0
#> 167  818983130 1297202769       CA(R)      1       0
#> 168  818983130 1141636389 MB(+ACA)(R)      1       0
#> 169  818983130 1141636389       MB(R)      1       0
#> 170  818983130 1141636389       CA(R)      1       0
#> 171  818983130  975432645       LH(R)      1       0
#> 172  818983130  974753207       LH(R)      1       0
#> 173  818983130  947453617       LH(R)      1       0
#> 174  818983130  944055652       LH(R)      1       0
#> 175  818983130  912022866       LH(R)      1       0
#> 176  818983130  851317951        None      1       0
#> 178  818983130  851317951     mALT(R)      1       0
#> 179  818983130  789248033       LH(R)      1       0
#> 180  818983130  760569606       LH(R)      1       0
#> 182  818983130  759810119       LH(R)      1       0
#> 186  818983130  665439889      SNP(R)      1       0
#> 187  818983130  665439889      SLP(R)      1       0
#> 188  818983130  667486204       LH(R)      1       0
#> 189  818983130  638481581       LH(R)      1       0
#> 191  818983130  635416407        None      1       0
#> 192  818983130  635416407     mALT(R)      1       0
#> 193  818983130  604726283       LH(R)      1       0
#> 195  818983130  570643102 MB(+ACA)(R)      1       0
#> 196  818983130  570643102       MB(R)      1       0
#> 197  818983130  570643102       CA(R)      1       0
#> 199  818983130  541957126       LH(R)      1       0
#> 200  818983130  541632990       LH(R)      1       0
#> 203  818983130  514406435       LH(R)      1       0
#> 204  818983130  487144598 MB(+ACA)(R)      1       0
#> 205  818983130  487144598       MB(R)      1       0
#> 207  818983130  487144598       CA(R)      1       0
#> 208  818983130  509695478 MB(+ACA)(R)      1       0
#> 209  818983130  509695478       MB(R)      1       0
#> 210  818983130  509695478       CA(R)      1       0
#> 212  818983130  481264051       LH(R)      1       0
#> 213  818983130  478613576       LH(R)      1       0
#> 217  818983130  356153135       LH(R)      1       0
#> 219 1796818119 7112613651       LH(R)      1       0
#> 221 1796818119 5901203780       LH(R)      1       0
#> 224 1796818119 5813087438       LH(R)      1       0
#> 226 1796818119 5813076969       AL(R)      1       0
#> 228 1796818119 5813069085       AL(R)      1       0
#> 233 1796818119 5813054725       AL(R)      1       0
#> 239 1796818119 5813047691       AL(R)      1       0
#> 243 1796818119 5813021291        None      1       0
#> 244 1796818119 5813021291     mALT(R)      1       0
#> 245 1796818119 5813009781       LH(R)      1       0
#> 247 1796818119 5812983673        None      1       0
#> 248 1796818119 5812983673     mALT(R)      1       0
#> 249 1796818119 5812982547        None      1       0
#> 250 1796818119 5812982547     mALT(R)      1       0
#> 252 1796818119 2040301572       AL(R)      1       0
#> 258 1796818119 1915837804       AL(R)      1       0
#> 259 1796818119 1915159759       AL(R)      1       0
#> 261 1796818119 1890613079       AL(R)      1       0
#> 262 1796818119 1890604495       AL(R)      1       0
#> 264 1796818119 1889922361       AL(R)      1       0
#> 268 1796818119 1856820145       AL(R)      1       0
#> 271 1796818119 1859574530       AL(R)      1       0
#> 272 1796818119 1859570213       AL(R)      1       0
#> 273 1796818119 1859570142       AL(R)      1       0
#> 277 1796818119 1859228714       AL(R)      1       0
#> 281 1796818119 1828544102       AL(R)      1       0
#> 285 1796818119 1859220337       AL(R)      1       0
#> 287 1796818119 1828203598       AL(R)      1       0
#> 292 1796818119 1827861966        None      1       0
#> 294 1796818119 1827848800       AL(R)      1       0
#> 297 1796818119 1827516355       AL(R)      1       0
#> 300 1796818119 1826121523       AL(R)      1       0
#> 302 1796818119 1825085656       AL(R)      1       0
#> 308 1796818119 1796817841       LH(R)      1       0
#> 312 1796818119 1796481518       AL(R)      1       0
#> 314 1796818119 1765778439       AL(R)      1       0
#> 315 1796818119 1763360713       AL(R)      1       0
#> 319 1796818119 1733387903       AL(R)      1       0
#> 328 1796818119 1702336197       AL(R)      1       0
#> 329 1796818119 1702306047       AL(R)      1       0
#> 331 1796818119 1702305987       AL(R)      1       0
#> 334 1796818119 1701757711       AL(R)      1       0
#> 336 1796818119 1699974843       AL(R)      1       0
#> 343 1796818119 1670265016 MB(+ACA)(R)      1       0
#> 344 1796818119 1670265016       MB(R)      1       0
#> 345 1796818119 1670265016       LH(R)      1       0
#> 346 1796818119 1670265016       CA(R)      1       0
#> 347 1796818119 1668275074       AL(R)      1       0
#> 348 1796818119 1669946117       AL(R)      1       0
#> 349 1796818119 1670934213       LH(R)      1       0
#> 351 1796818119 1669699714       AL(R)      1       0
#> 362 1796818119 1630051981        None      1       0
#> 363 1796818119 1630051981     mALT(R)      1       0
#> 364 1796818119 1640887603       AL(R)      1       0
#> 367 1796818119  974413307       LH(R)      1       0
#> 371 1796818119  917086906       LH(R)      1       0
#> 372 1796818119  917842566       LH(R)      1       0
#> 374 1796818119  883031393       LH(R)      1       0
#> 375 1796818119  881991445       LH(R)      1       0
#> 376 1796818119  851317951        None      1       0
#> 378 1796818119  851317951     mALT(R)      1       0
#> 379 1796818119  852976677       LH(R)      1       0
#> 381 1796818119  850260732       LH(R)      1       0
#> 382 1796818119  823999989       LH(R)      1       0
#> 385 1796818119  818983130       AL(R)      1       0
#> 386 1796818119  791303168       LH(R)      1       0
#> 387 1796818119  789248033       LH(R)      1       0
#> 388 1796818119  759586495       LH(R)      1       0
#> 389 1796818119  761938645       LH(R)      1       0
#> 391 1796818119  759810119       LH(R)      1       0
#> 394 1796818119  730537677       LH(R)      1       0
#> 398 1796818119  699031185        None      1       0
#> 399 1796818119  699031185     mALT(R)      1       0
#> 401 1796818119  698180393       LH(R)      1       0
#> 402 1796818119  667486204       LH(R)      1       0
#> 406 1796818119  611323175       LH(R)      1       0
#> 407 1796818119  605754137       LH(R)      1       0
#> 408 1796818119  574037266       AL(R)      1       0
#> 410 1796818119  548545143       LH(R)      1       0
#> 411 1796818119  542665377       LH(R)      1       0
#> 413 1796818119  539975122 MB(+ACA)(R)      1       0
#> 414 1796818119  539975122       MB(R)      1       0
#> 415 1796818119  539975122       CA(R)      1       0
#> 417 1796818119  512982243       LH(R)      1       0
#> 418 1796818119  487501700 MB(+ACA)(R)      1       0
#> 419 1796818119  487501700       MB(R)      1       0
#> 420 1796818119  487501700       CA(R)      1       0
#> 422 1796818119  480590566 MB(+ACA)(R)      1       0
#> 423 1796818119  480590566       MB(R)      1       0
#> 424 1796818119  480590566       CA(R)      1       0
#> 427 1796818119  478613576     mALT(R)      1       0
#> 429 1796818119  478613576        None      1       0
#> 431 1796818119  454693209       LH(R)      1       0
#> 434 1796818119  425790257     mALT(R)      1       0
#> 435 1796818119  425790257        None      1       0
#> 437 1796818119  359524536       LH(R)      1       0
#> 438 1796818119  328533761       LH(R)      1       0