Plot cosine clustering of neuprint/hemibrain neurons

Plot cosine clustering of neuprint/hemibrain neurons

neuprint_cosine_plot(
  x,
  partners = c("inputs", "outputs"),
  threshold = 5,
  method = c("ward.D", "single", "complete", "average", "mcquitty", "median", "centroid",
    "ward.D2"),
  labRow = "{type}",
  nas = c("zero", "drop"),
  conn = NULL,
  ...
)

Arguments

x

Query ids. May including searches in the style of neuprint_ids

partners

Whether to cluster based on connections to input or output partner neurons (default both).

threshold

An integer threshold (connections >= this will be returned)

method

The cluster method to use (see hclust)

labRow

A vector defining the row labels. If it is a character vector length 1 containing any curly braces it will be interpreted as a glue string for interpolation, using the data.frame returned by neuprint_get_meta as a source of information.

nas

What to do with entries that have NAs. Default is to set them to 0 similarity.

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

Additional arguments passed to heatmap

Value

The result of heatmap invisibly including the row and column dendrograms.

See also

Examples

# \donttest{
neuprint_cosine_matrix("/DA[1-3].*PN", grepl("ORN",type), partners='in')
#> Error in neuprint_cosine_matrix("/DA[1-3].*PN", grepl("ORN", type), partners = "in"): could not find function "neuprint_cosine_matrix"
# }
if (FALSE) {
# Considering all partners
neuprint_cosine_matrix("/lLN2.+", partners='in') %>% neuprint_cosine_plot()
# excluding the PN partners from the cosine distance score
neuprint_cosine_matrix("/lLN2.+", !grepl("PN",type), partners='in') %>% neuprint_cosine_plot()
# just use PN partners for cosine distance score
neuprint_cosine_matrix("/lLN2.+", grepl("PN",type), partners='in') %>% neuprint_cosine_plot()
}