Download data from neuprint for specified bodyids / a specified ROI

Download neuron morphology and connectivity data to a specified directrory as .csv and .rda files

neuprint_dump(dir, bodyids = NULL, roi = NULL, preprocess = NULL,
  connectivity = TRUE, volumes = TRUE, meta = TRUE, nat = TRUE,
  drvid = TRUE, flow.centrality = FALSE, soma = TRUE,
  estimate.soma = FALSE, heal = TRUE, connectors = TRUE,
  all_segments = TRUE, resample = FALSE, scale = 4,
  voxel.thresh = 1e+07, split = c("postsynapses", "presynapses",
  "distance"), dataset = NULL, conn = NULL, OmitFailures = TRUE, ...)

Arguments

dir
the directory to which to save the dump

bodyids
the body IDs for neurons/segments (bodies) you wish to query

roi
a single ROI. Use neuprint_ROIs to see what is available.

preprocess
a function that can be applied to a nat::neuronlist object, to be called on neurons once they are read from the neuprint server using neuprint_read_neurons

connectivity
whether or not to include connectivity information in the dump, i.e. an adjacency matrix between bodyids

volumes
whether or not to include neuron volume information in the dump, i.e. voxels. Currently only works by talking to a DVID server using the package drvid

meta
whether or not to fetch a meta data for the given bodyids, using neuprint_get_meta

nat
whether or not to read neurons are nat::neuronlist objects (TRUE) or get SWC data frame (FALSE)

drvid
whether or not to use drvid::read.neuron.dvid rather than a cypher post request to neuprint_fetch_custom, in order to read a neuron. This might be faster, and this might also enable access to skeletons on an underlying DVID database that have not yet been ported to neuprint.

flow.centrality
if TRUE, the neuron is split into primary neurite, axon and dendrite based on its flow centrality, as specified by Schneider-Mizell et al. (2016), using catnat::flow.centrality

soma
whether or not to fetch a possible soma location for the given bodyids, using neuprint_locate_soma

estimate.soma
if soma = TRUE, and estimate.soma = TRUE, then when a soma has not been tagged in rhe dataset, one is estimated by finding the leaf node with the largest mean geodesic distance from all synapses

heal
whether or not to heal a fragmented skeleton using a minimum spanning tree, via heal_skeleton

connectors
whether or not to add synapse data to the retrieved skeletons in the format used by the rcatmaid package, for easy use with rcatmaid or catnat functions. This can be done for synapse-less skeletons using neuprint_assign_connectors

all_segments
if TRUE, all bodies are considered, if FALSE, only 'Neurons', i.e. bodies with a status roughly traced status.

resample
if a number, the neuron is resampled using nat::resample, stepsize = resample. If 0 or FALSE (default), no resampling occurs.

scale
Resolution of sparse volume starting with 0 where each level beyond 0 has 1/2 resolution of previous level. "coarse" will return the volume in block coordinates.

voxel.thresh
the size threshold, in number of voxels, a neuron/segment must exceed, to be included in the dump, if read from an ROI

split
the algorithm will assign two main neurite compartments, which as per SWC format will be indicates as either axon (Label =2) or dendrite (Label = 3) in the returned objects, at neuron$d$Label. This assignment can be based which compartment contains the most postsynapses ("postsynapses") or presynapses ("presynapses"), or the Euclidean distance of its first branch point from the primary branch point (i.e. the first branch point from the soma) ("distance").

dataset
optional, a dataset you want to query. If NULL, the default specified by your R environ file is used. See neuprint_login for details.

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

OmitFailures
Whether to omit neurons for which FUN gives an error. The default value (NA) will result in nlapply stopping with an error message the moment there is an error. For other values, see details.

...
methods passed to neuprint_login

See also