Get available metadata fields for Neuron nodes

Get available metadata fields for Neuron nodes

neuprint_get_fields(
  possibleFields = c("bodyId", "pre", "post", "upstream", "downstream", "status",
    "statusLabel", "cropped", "instance", "name", "size", "type", "cellBodyFiber",
    "somaLocation", "somaRadius", "notes"),
  negateFields = FALSE,
  dataset = NULL,
  conn = NULL,
  ...
)

Arguments

possibleFields

: field names to choose from

negateFields

: Whether to include (FALSE, the default) or exclude possibleFields

dataset

optional, a dataset you want to query. If NULL, the default specified by your R environ file is used or, failing that the current connection, is used. See neuprint_login for details.

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

additional arguments passed to neuprint_fetch_custom

Value

a vector of available fields

Examples

# \donttest{
neuprint_get_fields()
#>  [1] "bodyId"        "pre"           "post"          "upstream"     
#>  [5] "downstream"    "status"        "statusLabel"   "size"         
#>  [9] "cropped"       "instance"      "type"          "cellBodyFiber"
#> [13] "somaLocation"  "somaRadius"    "notes"        
# }