Get a list of the shortest paths between two neurons

Get all of the shortest paths in the database that connect the query neurons with at least weightT synapses at each step

neuprint_get_shortest_paths(
  body_pre,
  body_post,
  weightT = 5,
  roi = NULL,
  by.roi = FALSE,
  chunk = TRUE,
  progress = FALSE,
  dataset = NULL,
  conn = NULL,
  all_segments = FALSE,
  ...
)

Arguments

body_pre

the bodyid of the neuron at the start of the path

body_post

the bodyid of the neuron at the end of the path

weightT

weight threshold

roi

Limit the search to connections happening within a certain ROI or set of ROIs (NULL by default)

by.roi

Return the results by ROI. Default to FALSE

chunk

A logical specifying whether to split the query into multiple chunks or an integer specifying the size of those chunks (which defaults to 5 when chunk=TRUE).

progress

if TRUE, a progress bar will be shown. This may slow the data fetching process for smaller queries. The default of progress=NULL will only show a progress bar if the query will be split into multiple chunks based on the chunk argument.

dataset

optional, a dataset you want to query. If NULL, the default specified by your R environ file is used or, failing that the current connection, is used. See neuprint_login for details.

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

all_segments

if TRUE, all bodies are considered, if FALSE, only 'Neurons', i.e. bodies with a status roughly traced status.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
neuprint_get_shortest_paths(c(1128092885,481121605),5813041365,weightT=20)
#>        from         to weight depth         name.from           name.to
#> 1 481121605  677182018     49     1 PFL1(PB12a)_L2_C3          FB2B_b_R
#> 2 677182018 5813041365     22     2          FB2B_b_R PFL1(PB12a)_L4_C5
#>   type.from type.to
#> 1      PFL1  FB2B_b
#> 2    FB2B_b    PFL1
# }