Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-VA6(L)"         "VES(R)"            "AL-DP1m(R)"       
#>   [4] "AL-VM7d(R)"        "lACA(R)"           "aL(R)"            
#>   [7] "AL-VA7l(R)"        "AL-V(R)"           "EBr5"             
#>  [10] "AL-DL3(R)"         "AL-DA4l(R)"        "AL-VA5(R)"        
#>  [13] "AOT(R)"            "b'L(R)"            "AL-DM2(L)"        
#>  [16] "GNG"               "AL-DA3(R)"         "BU(L)"            
#>  [19] "PB(R9)"            "AL-VM5d(R)"        "CX"               
#>  [22] "EPA(R)"            "AL-DM1(L)"         "AL-VM7v(L)"       
#>  [25] "IPS(R)"            "PB(L4)"            "PB(L3)"           
#>  [28] "PB(L8)"            "AL-DC1(R)"         "AL-VC1(R)"        
#>  [31] "AL-VP1l(R)"        "NO3(R)"            "FBl2"             
#>  [34] "SLP(R)"            "AL-VL2a(R)"        "GF(R)"            
#>  [37] "AL-DM6(L)"         "vACA(R)"           "SIP(R)"           
#>  [40] "AL-DM2(R)"         "AL-VA3(R)"         "MB(L)"            
#>  [43] "PVLP(R)"           "NO2(R)"            "AL-VL1(R)"        
#>  [46] "g1(R)"             "LAL(R)"            "SPS(L)"           
#>  [49] "AME(R)"            "CRE(R)"            "dACA(R)"          
#>  [52] "a'L(R)"            "EBr2r4"            "AL-DC2(L)"        
#>  [55] "PB(R6)"            "PB(L1)"            "VMNP"             
#>  [58] "FBl5"              "b1(R)"             "AL-VA4(R)"        
#>  [61] "LX(R)"             "AVLP(R)"           "AL-VA7m(R)"       
#>  [64] "AL-VP5(R)"         "SIP(L)"            "g5(R)"            
#>  [67] "AL-VC3l(R)"        "AL-VM2(R)"         "SMP(R)"           
#>  [70] "AL-DA4m(L)"        "MB(+ACA)(R)"       "AB(L)"            
#>  [73] "PB"                "FBl7"              "CA(R)"            
#>  [76] "AL-VC4(R)"         "PB(L5)"            "FB"               
#>  [79] "FBl3"              "EPA(L)"            "NO2(L)"           
#>  [82] "PED(R)"            "g4(R)"             "PB(R2)"           
#>  [85] "AL-DL2v(R)"        "AL-VP4(R)"         "PB(L2)"           
#>  [88] "LOP(R)"            "mALT(R)"           "AL-DM4(R)"        
#>  [91] "AL-DM6(R)"         "AL-VM7v(R)"        "GC"               
#>  [94] "bL(R)"             "AL-DL4(R)"         "hemibrain"        
#>  [97] "AL-DA3(L)"         "PB(R4)"            "CRE(-RUB)(L)"     
#> [100] "LX(L)"             "AL-D(R)"           "AL-VA6(R)"        
#> [103] "PB(L6)"            "LO(R)"             "a'2(R)"           
#> [106] "AL-VC3m(R)"        "LAL(-GA)(R)"       "gL(L)"            
#> [109] "AL(L)"             "a'1(R)"            "AL-DL4(L)"        
#> [112] "FBl9"              "AL-VP1d(R)"        "AL-DP1m(L)"       
#> [115] "ICL(L)"            "ICL(R)"            "AL-DA2(R)"        
#> [118] "AL-DM1(R)"         "AL-DM5(R)"         "AL-VM4(R)"        
#> [121] "ATL(L)"            "AOTU(R)"           "GA(R)"            
#> [124] "AL-DL5(R)"         "WED(R)"            "NO1(L)"           
#> [127] "AL-DC2(R)"         "a'3(R)"            "LAL(L)"           
#> [130] "a3(R)"             "AL-DA1(R)"         "AL-VC2(R)"        
#> [133] "PB(R8)"            "IB112"             "AL-DA4m(R)"       
#> [136] "PB(R5)"            "SMP(L)"            "EBr6"             
#> [139] "EB"                "AL-D(L)"           "AL-VM5v(R)"       
#> [142] "AL-VP3(R)"         "CRE(-ROB,-RUB)(R)" "PRW"              
#> [145] "ROB(R)"            "gL(R)"             "INP"              
#> [148] "b'1(R)"            "SCL(L)"            "POC"              
#> [151] "b'2(R)"            "PLP(R)"            "BU(R)"            
#> [154] "NO"                "PB(L7)"            "AL-DL1(R)"        
#> [157] "AL-DM3(R)"         "PB(R7)"            "FLA(R)"           
#> [160] "AL-DC4(L)"         "PB(R1)"            "EBr3pw"           
#> [163] "FBl6"              "EBr1"              "FBl4"             
#> [166] "CA(L)"             "AL-DP1l(R)"        "SAD"              
#> [169] "GOR(L)"            "AL-DM5(L)"         "SNP(L)"           
#> [172] "FB-column3"        "a'L(L)"            "AL-DC4(R)"        
#> [175] "AL-VA1d(R)"        "AL-DL5(L)"         "AL-DM4(L)"        
#> [178] "ATL(R)"            "aL(L)"             "b'L(L)"           
#> [181] "AL-VL2p(R)"        "NO(L)"             "IB"               
#> [184] "SPS(R)"            "AL-VM7d(L)"        "PB(L9)"           
#> [187] "FBl1"              "AB(R)"             "AL-VM3(R)"        
#> [190] "AL-DM3(L)"         "CRE(L)"            "AL-DL2d(R)"       
#> [193] "PENP"              "AL-DC1(L)"         "bL(L)"            
#> [196] "a1(R)"             "FBl8"              "AMMC"             
#> [199] "b2(R)"             "EBr3d"             "NO(R)"            
#> [202] "RUB(L)"            "OL(R)"             "VES(L)"           
#> [205] "AL-DA2(L)"         "MB(R)"             "mALT(L)"          
#> [208] "ME(R)"             "SNP(R)"            "NO3(L)"           
#> [211] "PB(R3)"            "RUB(R)"            "a2(R)"            
#> [214] "SAD(-AMMC)"        "GOR(R)"            "AL-VA2(R)"        
#> [217] "g2(R)"             "EBr3am"            "AL(R)"            
#> [220] "AL-VM1(R)"         "g3(R)"             "SCL(R)"           
#> [223] "VLNP(R)"           "AL-VC5(R)"         "NO1(R)"           
#> [226] "AL-VA1v(R)"        "AL-VP2(R)"         "LH(R)"            
#> [229] "CAN(R)"            "AL-DC3(R)"         "AL-VP1m(R)"       
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "ATL(R)"            "AME(R)"            "VLNP(R)"          
#>   [4] "AOTU(R)"           "g1(R)"             "SNP(R)"           
#>   [7] "EBr2r4"            "FBl8"              "a'3(R)"           
#>  [10] "MB(L)"             "ME(R)"             "FBl1"             
#>  [13] "GOR(L)"            "EB"                "ROB(R)"           
#>  [16] "AB(R)"             "PB(L3)"            "CRE(L)"           
#>  [19] "PB(L1)"            "BU(R)"             "NO2(L)"           
#>  [22] "aL(R)"             "b2(R)"             "EBr6"             
#>  [25] "EBr5"              "ICL(L)"            "SLP(R)"           
#>  [28] "AB(L)"             "EBr1"              "a'2(R)"           
#>  [31] "a'L(R)"            "FBl2"              "aL(L)"            
#>  [34] "bL(R)"             "SMP(R)"            "SCL(L)"           
#>  [37] "b'2(R)"            "PB(R7)"            "PB(R8)"           
#>  [40] "a'1(R)"            "CAN(R)"            "CRE(R)"           
#>  [43] "dACA(R)"           "PB(R1)"            "NO(R)"            
#>  [46] "FBl6"              "LAL(R)"            "MB(R)"            
#>  [49] "SMP(L)"            "lACA(R)"           "GC"               
#>  [52] "BU(L)"             "PLP(R)"            "FBl7"             
#>  [55] "RUB(L)"            "PENP"              "LAL(L)"           
#>  [58] "SPS(R)"            "PB"                "bL(L)"            
#>  [61] "GNG"               "a1(R)"             "WED(R)"           
#>  [64] "SIP(R)"            "NO"                "EPA(L)"           
#>  [67] "mALT(L)"           "PB(R6)"            "VMNP"             
#>  [70] "gL(L)"             "a'L(L)"            "vACA(R)"          
#>  [73] "mALT(R)"           "MB(+ACA)(R)"       "PB(R4)"           
#>  [76] "GF(R)"             "FBl5"              "NO3(R)"           
#>  [79] "AL(L)"             "CA(R)"             "hemibrain"        
#>  [82] "SIP(L)"            "EBr3am"            "ATL(L)"           
#>  [85] "CX"                "g3(R)"             "LAL(-GA)(R)"      
#>  [88] "b'L(R)"            "SNP(L)"            "RUB(R)"           
#>  [91] "ICL(R)"            "GOR(R)"            "NO(L)"            
#>  [94] "NO1(L)"            "EBr3d"             "EPA(R)"           
#>  [97] "PB(R9)"            "AVLP(R)"           "SCL(R)"           
#> [100] "PB(L7)"            "VES(L)"            "PB(R5)"           
#> [103] "AMMC"              "SAD"               "LH(R)"            
#> [106] "PB(R2)"            "IPS(R)"            "NO2(R)"           
#> [109] "FB"                "VES(R)"            "FBl3"             
#> [112] "FBl4"              "CRE(-ROB,-RUB)(R)" "a3(R)"            
#> [115] "AL(R)"             "PB(L5)"            "a2(R)"            
#> [118] "FLA(R)"            "g4(R)"             "PVLP(R)"          
#> [121] "PB(R3)"            "CA(L)"             "SAD(-AMMC)"       
#> [124] "LX(L)"             "INP"               "b'L(L)"           
#> [127] "b'1(R)"            "PB(L9)"            "g5(R)"            
#> [130] "FBl9"              "POC"               "AOT(R)"           
#> [133] "LO(R)"             "PB(L4)"            "OL(R)"            
#> [136] "GA(R)"             "PRW"               "PED(R)"           
#> [139] "PB(L8)"            "NO1(R)"            "PB(L2)"           
#> [142] "LOP(R)"            "LX(R)"             "gL(R)"            
#> [145] "EBr3pw"            "b1(R)"             "IB"               
#> [148] "g2(R)"             "SPS(L)"            "PB(L6)"           
#> [151] "NO3(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "b'L(R)"            "LX(R)"             "AL-VP1d(R)"       
#>   [4] "g1(R)"             "AL-DM1(R)"         "SLP(R)"           
#>   [7] "EBr3am"            "gL(L)"             "PB(R4)"           
#>  [10] "OL(R)"             "AL-DL5(R)"         "VLNP(R)"          
#>  [13] "SAD"               "AL-DC4(L)"         "PVLP(R)"          
#>  [16] "CA(R)"             "AL-VA3(R)"         "AL-VC1(R)"        
#>  [19] "AOT(R)"            "PENP"              "AL-DA4m(L)"       
#>  [22] "AL-DA2(R)"         "AL-DL3(R)"         "GOR(L)"           
#>  [25] "AL-DM1(L)"         "LO(R)"             "SPS(R)"           
#>  [28] "AL-DL2v(R)"        "AL-DM2(R)"         "AL-VC3m(R)"       
#>  [31] "AL-DM3(L)"         "PB(L6)"            "a3(R)"            
#>  [34] "AL-VA5(R)"         "AL-VM3(R)"         "MB(L)"            
#>  [37] "AL-VP1m(R)"        "NO2(L)"            "PB(R7)"           
#>  [40] "AL-VA7m(R)"        "FLA(R)"            "AL-DA3(R)"        
#>  [43] "AL-DC4(R)"         "AL-VA6(R)"         "CA(L)"            
#>  [46] "SIP(L)"            "AB(R)"             "AL-DM2(L)"        
#>  [49] "a'1(R)"            "AL-DM5(R)"         "a'2(R)"           
#>  [52] "ROB(R)"            "CRE(L)"            "bL(L)"            
#>  [55] "ATL(R)"            "a2(R)"             "PB(L9)"           
#>  [58] "PB(L7)"            "IPS(R)"            "AL-DL2d(R)"       
#>  [61] "AL-VL2a(R)"        "dACA(R)"           "AL-VM7v(L)"       
#>  [64] "AL-DA1(R)"         "AMMC"              "mALT(R)"          
#>  [67] "FBl5"              "PB(L5)"            "g3(R)"            
#>  [70] "AL-DC2(L)"         "NO3(L)"            "PB(R3)"           
#>  [73] "CRE(-ROB,-RUB)(R)" "EBr2r4"            "NO(R)"            
#>  [76] "PB(R6)"            "NO3(R)"            "AL(R)"            
#>  [79] "AL-V(R)"           "PB(L3)"            "SMP(R)"           
#>  [82] "SCL(R)"            "AL-VC3l(R)"        "AL-VM4(R)"        
#>  [85] "AL-VM5v(R)"        "AL-VP5(R)"         "AL-VP3(R)"        
#>  [88] "LOP(R)"            "NO1(L)"            "EBr3d"            
#>  [91] "AL-DL4(L)"         "LX(L)"             "b'L(L)"           
#>  [94] "VMNP"              "AL-VM7v(R)"        "aL(R)"            
#>  [97] "CAN(R)"            "LAL(R)"            "AL-VA2(R)"        
#> [100] "mALT(L)"           "AVLP(R)"           "SNP(L)"           
#> [103] "AL-VA4(R)"         "PLP(R)"            "GC"               
#> [106] "EPA(R)"            "AL-DA2(L)"         "FBl8"             
#> [109] "aL(L)"             "POC"               "ATL(L)"           
#> [112] "LAL(-GA)(R)"       "AL-DP1m(L)"        "AME(R)"           
#> [115] "AL-DM3(R)"         "PB(L8)"            "AL-VL1(R)"        
#> [118] "PED(R)"            "hemibrain"         "EBr5"             
#> [121] "FBl9"              "AL-VM7d(R)"        "AOTU(R)"          
#> [124] "EPA(L)"            "INP"               "AL-VP2(R)"        
#> [127] "vACA(R)"           "SNP(R)"            "AL-DC2(R)"        
#> [130] "b'1(R)"            "VES(L)"            "GOR(R)"           
#> [133] "FBl3"              "AL-DP1m(R)"        "VES(R)"           
#> [136] "bL(R)"             "AL-VL2p(R)"        "PB(R9)"           
#> [139] "FBl2"              "FBl7"              "RUB(L)"           
#> [142] "BU(L)"             "a'L(R)"            "AL-DM4(L)"        
#> [145] "NO(L)"             "AL-D(R)"           "AL-VC4(R)"        
#> [148] "AL-VP4(R)"         "a1(R)"             "PB(R8)"           
#> [151] "AL-DC1(L)"         "AL-VM7d(L)"        "AL-DA4m(R)"       
#> [154] "AL-VA1d(R)"        "AL(L)"             "AL-DA3(L)"        
#> [157] "FBl4"              "MB(R)"             "AL-DC1(R)"        
#> [160] "FB"                "EBr1"              "EB"               
#> [163] "AL-DM6(L)"         "b2(R)"             "PB(L4)"           
#> [166] "AL-VM1(R)"         "PRW"               "AL-DM5(L)"        
#> [169] "g5(R)"             "IB"                "lACA(R)"          
#> [172] "AL-D(L)"           "ICL(R)"            "PB"               
#> [175] "FB-column3"        "AL-VM5d(R)"        "AL-VM2(R)"        
#> [178] "AL-DC3(R)"         "PB(R2)"            "AL-VA7l(R)"       
#> [181] "PB(L2)"            "FBl6"              "MB(+ACA)(R)"      
#> [184] "EBr6"              "EBr3pw"            "AL-DM4(R)"        
#> [187] "NO1(R)"            "CX"                "gL(R)"            
#> [190] "PB(R1)"            "ICL(L)"            "AL-VA1v(R)"       
#> [193] "AB(L)"             "SIP(R)"            "SPS(L)"           
#> [196] "g4(R)"             "AL-DL4(R)"         "AL-VC2(R)"        
#> [199] "a'3(R)"            "b'2(R)"            "CRE(-RUB)(L)"     
#> [202] "a'L(L)"            "GF(R)"             "NO2(R)"           
#> [205] "g2(R)"             "SAD(-AMMC)"        "GNG"              
#> [208] "AL-DA4l(R)"        "AL-VP1l(R)"        "SCL(L)"           
#> [211] "GA(R)"             "AL-DL1(R)"         "AL-DP1l(R)"       
#> [214] "AL-DL5(L)"         "CRE(R)"            "BU(R)"            
#> [217] "AL-VA6(L)"         "b1(R)"             "AL-VC5(R)"        
#> [220] "PB(R5)"            "RUB(R)"            "SMP(L)"           
#> [223] "PB(L1)"            "FBl1"              "WED(R)"           
#> [226] "LH(R)"             "NO"                "ME(R)"            
#> [229] "AL-DM6(R)"         "LAL(L)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "NO3(R)"            "SCL(R)"            "ME(R)"            
#>   [4] "PB(R2)"            "AL-DA1(R)"         "PB(L3)"           
#>   [7] "ROB(R)"            "CRE(-RUB)(L)"      "GNG"              
#>  [10] "NO(L)"             "b'L(R)"            "AL-DM6(L)"        
#>  [13] "RUB(L)"            "AL-DC1(L)"         "AL-DP1l(R)"       
#>  [16] "PB(R7)"            "AL-VA6(R)"         "OL(R)"            
#>  [19] "AL-D(L)"           "AL-VA6(L)"         "b1(R)"            
#>  [22] "AL-VP1l(R)"        "AL-VP1d(R)"        "PLP(R)"           
#>  [25] "AOTU(R)"           "a'L(R)"            "PB(L4)"           
#>  [28] "EBr3am"            "PB(L5)"            "g3(R)"            
#>  [31] "PB(R9)"            "AL-DA3(R)"         "AOT(R)"           
#>  [34] "FB"                "a'3(R)"            "EBr3d"            
#>  [37] "gL(R)"             "FB-column3"        "AL-VM3(R)"        
#>  [40] "GC"                "AL-VP3(R)"         "AL-DM6(R)"        
#>  [43] "PB(R8)"            "ICL(L)"            "AL-DC4(R)"        
#>  [46] "AL-VM1(R)"         "PB(R5)"            "NO1(R)"           
#>  [49] "g1(R)"             "EBr1"              "SAD(-AMMC)"       
#>  [52] "FBl8"              "LAL(-GA)(R)"       "AL-VC3l(R)"       
#>  [55] "PB(L2)"            "PB(L8)"            "AL-DA4l(R)"       
#>  [58] "AL-VM7d(L)"        "MB(L)"             "SLP(R)"           
#>  [61] "AL(R)"             "PENP"              "FBl4"             
#>  [64] "AL-DA4m(L)"        "PB(L6)"            "AL-VA7l(R)"       
#>  [67] "VLNP(R)"           "POC"               "AB(L)"            
#>  [70] "AL-DA2(L)"         "a'1(R)"            "a'L(L)"           
#>  [73] "AL-DC1(R)"         "aL(R)"             "ATL(R)"           
#>  [76] "PB(R4)"            "PB"                "AL-DL5(L)"        
#>  [79] "bL(R)"             "a3(R)"             "a1(R)"            
#>  [82] "EPA(R)"            "g4(R)"             "AL-DM2(L)"        
#>  [85] "PVLP(R)"           "FBl5"              "b'1(R)"           
#>  [88] "PB(L1)"            "AL-DP1m(L)"        "SNP(L)"           
#>  [91] "AVLP(R)"           "AL-DA2(R)"         "AL-DL2d(R)"       
#>  [94] "AL-DM4(L)"         "EBr5"              "AL-DM2(R)"        
#>  [97] "AL-VC1(R)"         "AL-VC2(R)"         "AL-VM7d(R)"       
#> [100] "SMP(R)"            "a2(R)"             "AL-DM5(L)"        
#> [103] "NO2(R)"            "AL-VM5v(R)"        "AL(L)"            
#> [106] "VMNP"              "AL-DL3(R)"         "AL-D(R)"          
#> [109] "AL-VM4(R)"         "LAL(L)"            "g5(R)"            
#> [112] "EB"                "AL-DL4(L)"         "MB(+ACA)(R)"      
#> [115] "VES(L)"            "AL-DC3(R)"         "AL-DM5(R)"        
#> [118] "AL-VL2a(R)"        "WED(R)"            "AL-VP4(R)"        
#> [121] "AL-VP1m(R)"        "dACA(R)"           "AL-VM7v(L)"       
#> [124] "AL-VA1v(R)"        "hemibrain"         "PB(L7)"           
#> [127] "AL-VC5(R)"         "SCL(L)"            "vACA(R)"          
#> [130] "EBr2r4"            "b'L(L)"            "mALT(L)"          
#> [133] "AL-DA3(L)"         "AL-VL1(R)"         "AL-DM4(R)"        
#> [136] "bL(L)"             "FBl7"              "ICL(R)"           
#> [139] "LX(R)"             "gL(L)"             "GOR(R)"           
#> [142] "GOR(L)"            "a'2(R)"            "FBl9"             
#> [145] "IB"                "SMP(L)"            "SPS(L)"           
#> [148] "LAL(R)"            "AL-DM1(R)"         "PB(R6)"           
#> [151] "AL-DM1(L)"         "PB(R3)"            "AL-DM3(R)"        
#> [154] "AMMC"              "SPS(R)"            "GA(R)"            
#> [157] "PED(R)"            "NO1(L)"            "EBr6"             
#> [160] "GF(R)"             "AL-DC2(R)"         "AL-VA1d(R)"       
#> [163] "EBr3pw"            "AL-DL5(R)"         "AL-VM2(R)"        
#> [166] "CRE(R)"            "CA(R)"             "PRW"              
#> [169] "SIP(R)"            "AL-VA2(R)"         "AL-V(R)"          
#> [172] "CRE(-ROB,-RUB)(R)" "SAD"               "AME(R)"           
#> [175] "CAN(R)"            "EPA(L)"            "AL-VL2p(R)"       
#> [178] "g2(R)"             "mALT(R)"           "FLA(R)"           
#> [181] "NO2(L)"            "lACA(R)"           "BU(R)"            
#> [184] "AL-VP2(R)"         "SIP(L)"            "AB(R)"            
#> [187] "AL-DA4m(R)"        "AL-VM7v(R)"        "LH(R)"            
#> [190] "BU(L)"             "SNP(R)"            "NO3(L)"           
#> [193] "FBl6"              "b2(R)"             "CX"               
#> [196] "AL-DL2v(R)"        "LOP(R)"            "AL-DM3(L)"        
#> [199] "IPS(R)"            "AL-VC3m(R)"        "NO"               
#> [202] "aL(L)"             "LX(L)"             "AL-VA4(R)"        
#> [205] "AL-VA5(R)"         "AL-VM5d(R)"        "RUB(R)"           
#> [208] "MB(R)"             "PB(R1)"            "AL-DP1m(R)"       
#> [211] "AL-VA7m(R)"        "AL-DC2(L)"         "FBl3"             
#> [214] "ATL(L)"            "b'2(R)"            "INP"              
#> [217] "PB(L9)"            "CRE(L)"            "FBl1"             
#> [220] "NO(R)"             "AL-DL1(R)"         "AL-DL4(R)"        
#> [223] "AL-DC4(L)"         "FBl2"              "CA(L)"            
#> [226] "VES(R)"            "AL-VA3(R)"         "AL-VC4(R)"        
#> [229] "AL-VP5(R)"         "LO(R)"            
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method   path                                    
#> [1,] "POST"   "/api/raw/cypher/transaction/:id/cypher"
#> [2,] "GET"    "/login"                                
#> [3,] "PATCH"  "/api"                                  
#> [4,] "REPORT" "/api"                                  
#> [5,] "POST"   "/api/custom/custom"                    
#> [6,] "POST"   "/api/npexplorer/simpleconnections"     
#>      name                                                                                             
#> [1,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                         
#> [2,] "github.com/janelia-flyem/echo-secure.loginHandler"                                              
#> [3,] "github.com/labstack/echo/v4.glob..func1"                                                        
#> [4,] "github.com/labstack/echo/v4.glob..func1"                                                        
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"                  
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }