Get the upstream and downstream connectivity of a neuron

Get the upstream and downstream connectivity of a body, restricted to within an ROI if specified

neuprint_connection_table(
  bodyids,
  prepost = c("PRE", "POST"),
  roi = NULL,
  by.roi = FALSE,
  threshold = 1L,
  details = FALSE,
  superLevel = FALSE,
  progress = FALSE,
  dataset = NULL,
  chunk = TRUE,
  all_segments = FALSE,
  conn = NULL,
  ...
)

Arguments

bodyids
the body IDs for neurons/segments (bodies) you wish to query. This can be in any form understood by neuprint_ids.

prepost
PRE: look for partners presynaptic (i.e upstream inputs) or POST: postsynaptic (downstream outputs) to the given bodyids

roi
a single ROI. Use neuprint_ROIs to see what is available.

by.roi
logical, whether or not to break neurons' connectivity down by region of interest (ROI)

threshold
Only return partners >= to an integer value. Default of 1 returns all partners. This threshold will be applied to the ROI weight when the roi argument is specified, otherwise to the whole neuron.

details
When TRUE returns adds a name and type column for partners.

superLevel
When by.roi=TRUE, should we look at low-level ROIs (superLevel=FALSE) or only super-level ROIs (superLevel=TRUE). A super-level ROIs can contain multiple lower-level ROIs. If set to `NULL`, both are returned.

progress
default FALSE. If TRUE, the API is called separately for each neuron and you can assess its progress, if an error is thrown by any one bodyid, that bodyid is ignored

dataset
optional, a dataset you want to query. If NULL, the default specified by your R environ file is used or, failing that the current connection, is used. See neuprint_login for details.

chunk
A logical specifying whether to split the query into multiple chunks or an integer specifying the size of those chunks (which defaults to 20 when chunk=TRUE).

all_segments
if TRUE, all bodies are considered, if FALSE, only 'Neurons', i.e. bodies with a status roughly traced status.

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

Value

a data frame with one row for each unique combination of query neuron, partner neuron and (where specified) ROI.

bodyid query neuron identifier

partner neuron identifier

prepost 0 for downstream/output partners; 1 for upstream/input partners

weight total number of connections between the query and partner neuron.

. When by.roi=TRUE

roi the name of the ROI

ROIweight the number of connections within that ROI

Note that when by.roi=TRUE there may be multiple rows for each unique pair of neurons and connections may appear in more than one row when some ROIs are contained within larger ROIs. For example the CA(R) is contained within the MB(R) ROI. Therefore you cannot sum the ROIweight to match the total connection strength (which is available as the weight column).

See also

Examples

# \donttest{
## Unitary connection strengths between two chosen neurons,
## and their downstream targets
c1 = neuprint_connection_table(c(818983130, 1796818119), prepost = "POST")
head(c1)

#>       bodyid    partner prepost weight
#> 1 1796818119  632402333       1     26
#> 2  818983130  785914742       1     24
#> 3 1796818119 5813080766       1     24
#> 4  818983130 1143677310       1     23
#> 5  818983130 5813055865       1     23
#> 6  818983130  798869986       1     23

## The same connections broken down by ROI
c2 = neuprint_connection_table(c(818983130, 1796818119), prepost = "POST",
                               by.roi = TRUE)
head(c2)

#>       bodyid   partner prepost weight         roi ROIweight
#> 1 1796818119 632402333       1     26 MB(+ACA)(R)        26
#> 2 1796818119 632402333       1     26       MB(R)        26
#> 3 1796818119 632402333       1     26       CA(R)        26
#> 4  818983130 785914742       1     24 MB(+ACA)(R)        23
#> 5  818983130 785914742       1     24       MB(R)        23
#> 6  818983130 785914742       1     24     mALT(R)         1

## The same connections broken down by super-level ROIs only
c3 = neuprint_connection_table(c(818983130, 1796818119), prepost = "POST",
                               by.roi = TRUE, superLevel = TRUE)
nrow(c3)

#> [1] 641

nrow(c2)

#> [1] 854

## Find only the connections within a specific ROI
c4 = neuprint_connection_table(c(818983130, 1796818119), prepost = "POST",
                               by.roi = TRUE, roi = "LH(R)")

# }