Download data from neuprint for specified bodyids / a specified ROI

Download neuron morphology and connectivity data to a specified directory as .csv and .rda files

neuprint_dump(
  dir,
  bodyids = NULL,
  roi = NULL,
  preprocess = NULL,
  connectivity = TRUE,
  volumes = TRUE,
  meta = TRUE,
  nat = TRUE,
  drvid = FALSE,
  soma = TRUE,
  heal = TRUE,
  connectors = TRUE,
  all_segments = FALSE,
  resample = FALSE,
  scale = 4,
  voxel.thresh = 1e+07,
  dataset = NULL,
  conn = NULL,
  OmitFailures = TRUE,
  ...
)

Arguments

dir

the directory to which to save the dump

bodyids

the body IDs for neurons/segments (bodies) you wish to query. This can be in any form understood by neuprint_ids.

roi

a single ROI. Use neuprint_ROIs to see what is available.

preprocess

a function that can be applied to a nat::neuronlist object, to be called on neurons once they are read from the neuprint server using neuprint_read_neurons

connectivity

whether or not to include connectivity information in the dump, i.e. an adjacency matrix between bodyids

volumes

whether or not to include neuron volume information in the dump, i.e. voxels. Currently only works by talking to a DVID server using the package drvid

meta

whether or not to fetch a meta data for the given bodyids, using neuprint_get_meta

nat

whether or not to read neurons are nat::neuronlist objects (TRUE) or get SWC data frame (FALSE)

drvid

whether or not to use drvid::read.neuron.dvid rather than a cypher post request to neuprint_fetch_custom, in order to read a neuron. This might be faster, and this might also enable access to skeletons on an underlying DVID database that have not yet been ported to neuprint.

soma

whether or not to fetch a possible soma location for the given bodyids, using neuprint_locate_soma

heal

whether or not to heal a fragmented skeleton using a minimum spanning tree, via nat::stitch_neurons_mst

connectors

whether or not to add synapse data to the retrieved skeletons in the format used by the rcatmaid package, for easy use with rcatmaid or catnat functions. This can be done for synapse-less skeletons using neuprint_assign_connectors

all_segments

if TRUE, all bodies are considered, if FALSE, only 'Neurons', i.e. bodies with a status roughly traced status.

resample

if a number, the neuron is resampled using nat::resample, stepsize = resample. If 0 or FALSE (default), no resampling occurs.

scale

Resolution of sparse volume starting with 0 where each level beyond 0 has 1/2 resolution of previous level. "coarse" will return the volume in block coordinates.

voxel.thresh

the size threshold, in number of voxels, a neuron/segment must exceed, to be included in the dump, if read from an ROI

dataset

optional, a dataset you want to query. If NULL, the default specified by your R environ file is used or, failing that the current connection, is used. See neuprint_login for details.

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

OmitFailures

Whether to omit neurons for which FUN gives an error. The default value (NA) will result in nlapply stopping with an error message the moment there is an error. For other values, see details.

...

methods passed to neuprint_login

See also