Get key metadata for body (including name, type, status, size)

Get key metadata for body (including name, type, status, size)

neuprint_get_meta(
  bodyids,
  dataset = NULL,
  all_segments = TRUE,
  conn = NULL,
  chunk = TRUE,
  progress = FALSE,
  possibleFields = NULL,
  ...
)

Arguments

bodyids
the body IDs for neurons/segments (bodies) you wish to query. This can be in any form understood by neuprint_ids.

dataset
optional, a dataset you want to query. If NULL, the default specified by your R environ file is used or, failing that the current connection, is used. See neuprint_login for details.

all_segments
if TRUE, all bodies are considered, if FALSE, only 'Neurons', i.e. bodies with a status roughly traced status.

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

chunk
A logical specifying whether to split the query into multiple chunks or an integer specifying the size of those chunks (which defaults to 2000 when chunk=TRUE).

progress
default FALSE. If TRUE, the API is called separately for each neuron and you can assess its progress, if an error is thrown by any one bodyid, that bodyid is ignored.

possibleFields
passed to neuprint_get_fields when not NULL, otherwise a default set are used.

...
methods passed to neuprint_login

Value

a data.frame containing the neuron's

name

type Cell type of the neuron

status (Traced etc)

statusLabel similar to status but often a bit more specific

size size in voxels

pre number of presynapses

post number of postsynapses

cropped whether the neuron is cropped by the hemibrain volume

soma whether the neuron has a soma in the hemibrain volume

cellBodyFiber names the tract connecting the soma to rest of neuron

Details

Sometimes a cellBodyFiber will be present even when the soma is not, so this may be a good test of if the neuron is present in the volume. The cellBodyFiber should be matched to (hemi)lineage information .

Examples

# \donttest{
da2s=neuprint_search(".*DA2.*")
neuprint_get_meta(da2s$bodyid)
#>       bodyid pre post status    statusLabel cropped    voxels              name
#> 1 1796817841 392  512 Traced Roughly traced   FALSE 723036360 mPN1_DA2(AVM02)_R
#> 2 1827516355 391  548 Traced Roughly traced   FALSE 737014743 mPN1_DA2(AVM02)_R
#> 3  818983130 408  562 Traced Roughly traced   FALSE 606338969 mPN1_DA2(AVM02)_R
#> 4 1797505019 339  479 Traced Roughly traced   FALSE 445320106 mPN1_DA2(AVM02)_R
#> 5 1796818119 510  818 Traced Roughly traced   FALSE 907369420 mPN1_DA2(AVM02)_R
#>      type cellBodyFiber  soma
#> 1 DA2 lPN           A12  TRUE
#> 2 DA2 lPN           A12  TRUE
#> 3 DA2 lPN           A12 FALSE
#> 4 DA2 lPN           A12 FALSE
#> 5 DA2 lPN           A12  TRUE
# or simpler
neuprint_get_meta('DA2')
#>       bodyid pre post status    statusLabel cropped    voxels              name
#> 1 1796817841 392  512 Traced Roughly traced   FALSE 723036360 mPN1_DA2(AVM02)_R
#> 2 1827516355 391  548 Traced Roughly traced   FALSE 737014743 mPN1_DA2(AVM02)_R
#> 3  818983130 408  562 Traced Roughly traced   FALSE 606338969 mPN1_DA2(AVM02)_R
#> 4 1797505019 339  479 Traced Roughly traced   FALSE 445320106 mPN1_DA2(AVM02)_R
#> 5 1796818119 510  818 Traced Roughly traced   FALSE 907369420 mPN1_DA2(AVM02)_R
#>      type cellBodyFiber  soma
#> 1 DA2 lPN           A12  TRUE
#> 2 DA2 lPN           A12  TRUE
#> 3 DA2 lPN           A12 FALSE
#> 4 DA2 lPN           A12 FALSE
#> 5 DA2 lPN           A12  TRUE
# }