Get key metadata for body (including name, type, status, size)

Get key metadata for body (including name, type, status, size)

neuprint_get_meta(
  bodyids,
  dataset = NULL,
  all_segments = TRUE,
  conn = NULL,
  chunk = TRUE,
  progress = FALSE,
  possibleFields = NULL,
  ...
)

Arguments

bodyids

the body IDs for neurons/segments (bodies) you wish to query. This can be in any form understood by neuprint_ids.

dataset

optional, a dataset you want to query. If NULL, the default specified by your R environ file is used or, failing that the current connection, is used. See neuprint_login for details.

all_segments

if TRUE, all bodies are considered, if FALSE, only 'Neurons', i.e. bodies with a status roughly traced status.

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

chunk

A logical specifying whether to split the query into multiple chunks or an integer specifying the size of those chunks (which defaults to 2000 when chunk=TRUE).

progress

default FALSE. If TRUE, the API is called separately for each neuron and you can assess its progress, if an error is thrown by any one bodyid, that bodyid is ignored.

possibleFields

passed to neuprint_get_fields when not NULL, otherwise a default set are used.

...

methods passed to neuprint_login

Value

a data.frame containing the neuron's

name

type Cell type of the neuron

status (Traced etc)

statusLabel similar to status but often a bit more specific

size size in voxels

pre number of presynapses

post number of postsynapses

cropped whether the neuron is cropped by the hemibrain volume

soma whether the neuron has a soma in the hemibrain volume

cellBodyFiber names the tract connecting the soma to rest of neuron

Details

Sometimes a cellBodyFiber will be present even when the soma is not, so this may be a good test of if the neuron is present in the volume. The cellBodyFiber should be matched to (hemi)lineage information .

Examples

# \donttest{
da2s=neuprint_search(".*DA2.*")
neuprint_get_meta(da2s$bodyid)
#>        bodyid pre post upstream downstream status    statusLabel    voxels
#> 1  1796817841 396  509      509       3275 Traced Roughly traced 726508881
#> 2  1796818119 511  818      818       4111 Traced Roughly traced 917547898
#> 3  1797505019 345  476      476       2932 Traced Roughly traced 450568345
#> 4  1827516355 391  548      548       3263 Traced Roughly traced 741928779
#> 5   818983130 409  562      562       3400 Traced Roughly traced 612641685
#> 6  1669017843  89  106      106        490 Traced         Leaves  83607944
#> 7  1669699714  89   95       95        482 Traced         Leaves  78128261
#> 8  1669946117 102  118      118        521 Traced         Leaves  99214765
#> 9  1670382036  78  102      102        412 Traced         Leaves  71552951
#> 10 1670386270  93   94       94        463 Traced         Leaves  82834198
#> 11 1670390647  91  127      127        485 Traced         Leaves  71396028
#> 12 1669358817  86  107      107        443 Traced         Leaves  76911124
#> 13 1671659809  18   17       17         66 Traced         Leaves   7008274
#> 14 1702107389  70   86       86        324 Traced         Leaves  71342150
#> 15 1702802179 103  110      110        522 Traced         Leaves  76354600
#> 16 1701675592 124  156      156        646 Traced         Leaves 102503828
#> 17 1701757711  88   86       86        433 Traced         Leaves  85122273
#> 18 1700393451  93  137      137        492 Traced         Leaves  74038623
#> 19 1732792926  86  114      114        518 Traced         Leaves  90045017
#> 20 1732797061  85   87       87        450 Traced         Leaves  78823995
#> 21 1733491818  97  109      109        499 Traced         Leaves  85400235
#> 22 1733833296 109  141      141        606 Traced         Leaves  90844980
#> 23 1796809680  67   69       69        317 Traced         Leaves  31604606
#> 24 1796813992  57   63       63        296 Traced         Leaves  26856261
#> 25 1796818061  83   95       95        451 Traced         Leaves  67833130
#> 26 1796818191  48   69       69        234 Traced         Leaves  23328521
#> 27 1796822470  96   98       98        496 Traced         Leaves  80017616
#> 28 1826929327 101  126      126        553 Traced         Leaves  89671264
#> 29 1858887650  43   54       54        210 Traced         Leaves  19559580
#> 30 1858892359  51   58       58        229 Traced         Leaves  26021432
#> 31 1859220826  41   35       35        205 Traced         Leaves  26734869
#> 32 1859224592  59   67       67        244 Traced         Leaves  31014767
#> 33 1859224739  50   44       44        233 Traced         Leaves  18505661
#> 34 1859565686  38   36       36        181 Traced         Leaves  16301601
#> 35 1889931525  50   50       50        237 Traced         Leaves  19397298
#> 36 5901198995  90  108      108        514 Traced         Leaves  74962468
#> 37 5901201379  21    2        2        112 Traced         Leaves  11362150
#> 38 5901202409 104  128      128        506 Traced         Leaves  80877299
#> 39 5813040847  81   77       77        415 Traced         Leaves  66608639
#> 40 5812994702  86  102      102        447 Traced         Leaves  90239045
#>    cropped      name    type cellBodyFiber notes  soma
#> 1    FALSE DA2_lPN_R DA2_lPN         AVM02  <NA>  TRUE
#> 2    FALSE DA2_lPN_R DA2_lPN         AVM02  <NA>  TRUE
#> 3    FALSE DA2_lPN_R DA2_lPN         AVM02  <NA> FALSE
#> 4    FALSE DA2_lPN_R DA2_lPN         AVM02  <NA>  TRUE
#> 5    FALSE DA2_lPN_R DA2_lPN         AVM02  <NA> FALSE
#> 6     TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 7     TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 8     TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 9     TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 10    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 11    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 12    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 13    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 14    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 15    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 16    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 17    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 18    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 19    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 20    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 21    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 22    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 23    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 24    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 25    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 26    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 27    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 28    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 29    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 30    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 31    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 32    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 33    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 34    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 35    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 36    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 37    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 38    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 39    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 40    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
# or simpler
neuprint_get_meta('DA2')
#>        bodyid pre post upstream downstream status    statusLabel    voxels
#> 1  1796817841 396  509      509       3275 Traced Roughly traced 726508881
#> 2  1796818119 511  818      818       4111 Traced Roughly traced 917547898
#> 3  1797505019 345  476      476       2932 Traced Roughly traced 450568345
#> 4  1827516355 391  548      548       3263 Traced Roughly traced 741928779
#> 5   818983130 409  562      562       3400 Traced Roughly traced 612641685
#> 6  1669017843  89  106      106        490 Traced         Leaves  83607944
#> 7  1669699714  89   95       95        482 Traced         Leaves  78128261
#> 8  1669946117 102  118      118        521 Traced         Leaves  99214765
#> 9  1670382036  78  102      102        412 Traced         Leaves  71552951
#> 10 1670386270  93   94       94        463 Traced         Leaves  82834198
#> 11 1670390647  91  127      127        485 Traced         Leaves  71396028
#> 12 1669358817  86  107      107        443 Traced         Leaves  76911124
#> 13 1671659809  18   17       17         66 Traced         Leaves   7008274
#> 14 1702107389  70   86       86        324 Traced         Leaves  71342150
#> 15 1702802179 103  110      110        522 Traced         Leaves  76354600
#> 16 1701675592 124  156      156        646 Traced         Leaves 102503828
#> 17 1701757711  88   86       86        433 Traced         Leaves  85122273
#> 18 1700393451  93  137      137        492 Traced         Leaves  74038623
#> 19 1732792926  86  114      114        518 Traced         Leaves  90045017
#> 20 1732797061  85   87       87        450 Traced         Leaves  78823995
#> 21 1733491818  97  109      109        499 Traced         Leaves  85400235
#> 22 1733833296 109  141      141        606 Traced         Leaves  90844980
#> 23 1796809680  67   69       69        317 Traced         Leaves  31604606
#> 24 1796813992  57   63       63        296 Traced         Leaves  26856261
#> 25 1796818061  83   95       95        451 Traced         Leaves  67833130
#> 26 1796818191  48   69       69        234 Traced         Leaves  23328521
#> 27 1796822470  96   98       98        496 Traced         Leaves  80017616
#> 28 1826929327 101  126      126        553 Traced         Leaves  89671264
#> 29 1858887650  43   54       54        210 Traced         Leaves  19559580
#> 30 1858892359  51   58       58        229 Traced         Leaves  26021432
#> 31 1859220826  41   35       35        205 Traced         Leaves  26734869
#> 32 1859224592  59   67       67        244 Traced         Leaves  31014767
#> 33 1859224739  50   44       44        233 Traced         Leaves  18505661
#> 34 1859565686  38   36       36        181 Traced         Leaves  16301601
#> 35 1889931525  50   50       50        237 Traced         Leaves  19397298
#> 36 5901198995  90  108      108        514 Traced         Leaves  74962468
#> 37 5901201379  21    2        2        112 Traced         Leaves  11362150
#> 38 5901202409 104  128      128        506 Traced         Leaves  80877299
#> 39 5813040847  81   77       77        415 Traced         Leaves  66608639
#> 40 5812994702  86  102      102        447 Traced         Leaves  90239045
#>    cropped      name    type cellBodyFiber notes  soma
#> 1    FALSE DA2_lPN_R DA2_lPN         AVM02  <NA>  TRUE
#> 2    FALSE DA2_lPN_R DA2_lPN         AVM02  <NA>  TRUE
#> 3    FALSE DA2_lPN_R DA2_lPN         AVM02  <NA> FALSE
#> 4    FALSE DA2_lPN_R DA2_lPN         AVM02  <NA>  TRUE
#> 5    FALSE DA2_lPN_R DA2_lPN         AVM02  <NA> FALSE
#> 6     TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 7     TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 8     TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 9     TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 10    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 11    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 12    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 13    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 14    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 15    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 16    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 17    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 18    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 19    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 20    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 21    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 22    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 23    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 24    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 25    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 26    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 27    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 28    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 29    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 30    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 31    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 32    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 33    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 34    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 35    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 36    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 37    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 38    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 39    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
#> 40    TRUE   ORN_DA2 ORN_DA2          <NA>  <NA> FALSE
# }
if (FALSE) {
neuprint_get_meta('cropped:false')
}