Get roiInfo associated with a body

Return pre and post counts in all the ROIs given bodyids innervate.

neuprint_get_roiInfo(
  bodyids,
  dataset = NULL,
  all_segments = TRUE,
  conn = NULL,
  ...
)

Arguments

bodyids
the body IDs for neurons/segments (bodies) you wish to query

dataset
optional, a dataset you want to query. If NULL, the default specified by your R environ file is used. See neuprint_login for details.

all_segments
if TRUE, all bodies are considered, if FALSE, only 'Neurons', i.e. bodies with a status roughly traced status.

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

Value

a dataframe, one row for each given body id, columns ROI_pre and ROI_post for every ROI. If data is missing, NA is returned.

Examples

neuprint_get_roiInfo(c(818983130, 1796818119))

#>       bodyid LH(R).pre LH(R).post MB(R).pre MB(R).post CA(R).pre CA(R).post
#> 1  818983130       210         97        75         14        75         14
#> 2 1796818119       283        117        77         16        77         16
#>   MB(+ACA)(R).pre MB(+ACA)(R).post SNP(R).pre SNP(R).post SLP(R).pre
#> 1              75               14          5           1          5
#> 2              77               16         NA          NA         NA
#>   SLP(R).post mALT(R).pre mALT(R).post AL(R).pre AL(R).post None.pre None.post
#> 1           1           3            6       113        442        2         2
#> 2          NA          13           18       136        666        1         1