Get roiInfo associated with a body

Return pre and post counts in all the ROIs given bodyids innervate.

neuprint_get_roiInfo(
  bodyids,
  dataset = NULL,
  all_segments = FALSE,
  chunk = TRUE,
  progress = FALSE,
  conn = NULL,
  ...
)

Arguments

bodyids

the body IDs for neurons/segments (bodies) you wish to query. This can be in any form understood by neuprint_ids.

dataset

optional, a dataset you want to query. If NULL, the default specified by your R environ file is used or, failing that the current connection, is used. See neuprint_login for details.

all_segments

if TRUE, all bodies are considered, if FALSE, only 'Neurons', i.e. bodies with a status roughly traced status.

chunk

A logical specifying whether to split the query into multiple chunks or an integer specifying the size of those chunks (which defaults to 2000 when chunk=TRUE).

progress

default FALSE. If TRUE, the API is called separately for each neuron and you can assess its progress, if an error is thrown by any one bodyid, that bodyid is ignored

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

Value

a dataframe, one row for each given body id, columns ROI_pre and ROI_post for every ROI. If data is missing, NA is returned.

Examples

# \donttest{
neuprint_get_roiInfo(c(818983130, 1796818119))
#> Error in neuprint_login(conn): Sorry you must specify a neuprint server! See ?neuprint_login for details!
# }