Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0`
#> $`hemibrain:v1.0`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0`$ROIs
#>   [1] "PB(R6)"            "lACA(R)"           "ME(R)"            
#>   [4] "gL(L)"             "PB(L7)"            "FBl4"             
#>   [7] "AVLP(R)"           "RUB(L)"            "ATL(L)"           
#>  [10] "b'L(R)"            "EBr6"              "AL(L)"            
#>  [13] "PB(L5)"            "NO1(L)"            "SPS(L)"           
#>  [16] "EPA(L)"            "CX"                "NO2(L)"           
#>  [19] "g3(R)"             "SAD"               "ICL(L)"           
#>  [22] "FBl9"              "AME(R)"            "NO2(R)"           
#>  [25] "LAL(R)"            "SCL(L)"            "LAL(-GA)(R)"      
#>  [28] "PB"                "EBr2r4"            "CRE(-ROB,-RUB)(R)"
#>  [31] "FBl2"              "FB"                "EB"               
#>  [34] "OL(R)"             "mALT(L)"           "GC"               
#>  [37] "PB(R9)"            "g5(R)"             "ICL(R)"           
#>  [40] "SLP(R)"            "PB(L2)"            "PB(L8)"           
#>  [43] "SNP(R)"            "PB(L6)"            "CRE(L)"           
#>  [46] "aL(L)"             "IPS(R)"            "AL(R)"            
#>  [49] "AOTU(R)"           "g1(R)"             "NO(R)"            
#>  [52] "gL(R)"             "FBl6"              "FBl7"             
#>  [55] "EBr3am"            "g4(R)"             "FBl3"             
#>  [58] "GOR(L)"            "SAD(-AMMC)"        "PB(R3)"           
#>  [61] "g2(R)"             "SCL(R)"            "LX(R)"            
#>  [64] "EPA(R)"            "PB(R1)"            "CA(R)"            
#>  [67] "IB"                "b'2(R)"            "b2(R)"            
#>  [70] "VMNP"              "a'2(R)"            "PB(R8)"           
#>  [73] "EBr3d"             "CAN(R)"            "GA(R)"            
#>  [76] "PLP(R)"            "FBl5"              "PRW"              
#>  [79] "CRE(R)"            "SIP(R)"            "VES(L)"           
#>  [82] "EBr5"              "PB(R2)"            "b1(R)"            
#>  [85] "b'1(R)"            "a'3(R)"            "NO3(L)"           
#>  [88] "AB(R)"             "GOR(R)"            "SMP(R)"           
#>  [91] "MB(+ACA)(R)"       "ATL(R)"            "PB(R4)"           
#>  [94] "bL(R)"             "LOP(R)"            "CA(L)"            
#>  [97] "WED(R)"            "INP"               "PB(L4)"           
#> [100] "LAL(L)"            "NO1(R)"            "NO"               
#> [103] "LO(R)"             "FBl8"              "VLNP(R)"          
#> [106] "POC"               "PB(R5)"            "BU(R)"            
#> [109] "PB(L9)"            "a3(R)"             "RUB(R)"           
#> [112] "SPS(R)"            "EBr1"              "PB(L1)"           
#> [115] "a'1(R)"            "EBr3pw"            "FBl1"             
#> [118] "AB(L)"             "bL(L)"             "LH(R)"            
#> [121] "b'L(L)"            "SIP(L)"            "PB(R7)"           
#> [124] "PVLP(R)"           "PED(R)"            "mALT(R)"          
#> [127] "NO3(R)"            "MB(L)"             "BU(L)"            
#> [130] "FLA(R)"            "PENP"              "GF(R)"            
#> [133] "SMP(L)"            "PB(L3)"            "hemibrain"        
#> [136] "a'L(R)"            "vACA(R)"           "a1(R)"            
#> [139] "a2(R)"             "aL(R)"             "SNP(L)"           
#> [142] "GNG"               "NO(L)"             "MB(R)"            
#> [145] "a'L(L)"            "dACA(R)"           "ROB(R)"           
#> [148] "VES(R)"            "LX(L)"             "AOT(R)"           
#> [151] "AMMC"             
#> 
#> $`hemibrain:v1.0`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method   path                                          
#> [1,] "GET"    "/api/v:ver/npexplorer/distribution"          
#> [2,] "HEAD"   "/api"                                        
#> [3,] "DELETE" "/api/*"                                      
#> [4,] "POST"   "/api/npexplorer/roisinneuron"                
#> [5,] "GET"    "/api/v:ver/npexplorer/commonconnectivity"    
#> [6,] "POST"   "/api/v:ver/raw/cypher/transaction/:id/commit"
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [2,] "github.com/labstack/echo.glob..func1"                                                      
#> [3,] "github.com/labstack/echo.glob..func1"                                                      
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }