Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "CRE(R)"            "AL-DL2v(R)"        "a'L(L)"           
#>   [4] "AL-DA3(L)"         "AL-VC3m(R)"        "vACA(R)"          
#>   [7] "NO2(L)"            "EBr2r4"            "AL-DL4(L)"        
#>  [10] "LO(R)"             "AL-DM2(R)"         "AL-VC4(R)"        
#>  [13] "AL-VM4(R)"         "g3(R)"             "NO(R)"            
#>  [16] "GNG"               "AL-DP1m(R)"        "IB112"            
#>  [19] "OL(R)"             "AL-DM4(R)"         "AL-VA3(R)"        
#>  [22] "PB(L5)"            "SMP(L)"            "a2(R)"            
#>  [25] "AL(R)"             "EBr3d"             "LH(R)"            
#>  [28] "AL-DC4(R)"         "RUB(R)"            "PB(R6)"           
#>  [31] "b'2(R)"            "AL-VM1(R)"         "AL-VM5d(R)"       
#>  [34] "SCL(L)"            "SCL(R)"            "AL(L)"            
#>  [37] "aL(L)"             "AL-VA4(R)"         "BU(L)"            
#>  [40] "AL-VA1d(R)"        "NO1(R)"            "CRE(L)"           
#>  [43] "FB-column3"        "FBl9"              "PB(R3)"           
#>  [46] "SAD"               "BU(R)"             "AL-VP1l(R)"       
#>  [49] "FBl7"              "AL-VC5(R)"         "INP"              
#>  [52] "POC"               "a'3(R)"            "AMMC"             
#>  [55] "GC"                "SAD(-AMMC)"        "AL-DC1(L)"        
#>  [58] "AL-DL5(L)"         "AL-DM1(L)"         "FBl1"             
#>  [61] "LX(L)"             "AL-D(R)"           "mALT(R)"          
#>  [64] "g1(R)"             "MB(+ACA)(R)"       "SLP(R)"           
#>  [67] "AL-DM6(R)"         "AL-VC1(R)"         "VLNP(R)"          
#>  [70] "LOP(R)"            "AL-DM4(L)"         "bL(L)"            
#>  [73] "AME(R)"            "AL-DL2d(R)"        "AL-VA5(R)"        
#>  [76] "EB"                "PVLP(R)"           "b'1(R)"           
#>  [79] "AL-VA7m(R)"        "PB(L3)"            "FBl5"             
#>  [82] "ICL(R)"            "b1(R)"             "mALT(L)"          
#>  [85] "VMNP"              "AL-VM7d(L)"        "FBl3"             
#>  [88] "PB(L2)"            "AL-VP3(R)"         "AL-VA6(R)"        
#>  [91] "AL-VM3(R)"         "CAN(R)"            "AL-DL5(R)"        
#>  [94] "AL-VP5(R)"         "AB(L)"             "SIP(L)"           
#>  [97] "a'L(R)"            "PB(L6)"            "AL-DC3(R)"        
#> [100] "AL-VP2(R)"         "FB"                "SNP(R)"           
#> [103] "PB(L8)"            "FBl8"              "NO(L)"            
#> [106] "CA(L)"             "SIP(R)"            "PB(R2)"           
#> [109] "b'L(L)"            "AL-VM7v(R)"        "hemibrain"        
#> [112] "dACA(R)"           "PB(R4)"            "AL-VM5v(R)"       
#> [115] "AL-VM7d(R)"        "b'L(R)"            "AL-DL1(R)"        
#> [118] "AL-DA2(R)"         "NO"                "VES(L)"           
#> [121] "EBr3am"            "PB(L7)"            "AL-VC2(R)"        
#> [124] "b2(R)"             "AL-VP1d(R)"        "AL-DA1(R)"        
#> [127] "AL-DM3(R)"         "lACA(R)"           "WED(R)"           
#> [130] "AL-DM3(L)"         "AL-VA2(R)"         "SMP(R)"           
#> [133] "a1(R)"             "AOTU(R)"           "SPS(R)"           
#> [136] "AL-VP4(R)"         "EBr5"              "RUB(L)"           
#> [139] "LX(R)"             "a3(R)"             "GOR(L)"           
#> [142] "AL-VL2a(R)"        "EBr3pw"            "AL-DM1(R)"        
#> [145] "ATL(L)"            "gL(L)"             "PB(R1)"           
#> [148] "AL-DP1m(L)"        "NO2(R)"            "GOR(R)"           
#> [151] "EPA(R)"            "AL-DM2(L)"         "a'2(R)"           
#> [154] "AL-DC2(L)"         "AL-VP1m(R)"        "MB(L)"            
#> [157] "PB(L1)"            "CA(R)"             "AL-DA4m(R)"       
#> [160] "AL-VM2(R)"         "CRE(-ROB,-RUB)(R)" "NO1(L)"           
#> [163] "g4(R)"             "AL-D(L)"           "gL(R)"            
#> [166] "AL-DC2(R)"         "AL-DL3(R)"         "PED(R)"           
#> [169] "PB(R8)"            "CRE(-RUB)(L)"      "EBr1"             
#> [172] "PLP(R)"            "LAL(R)"            "AL-VA7l(R)"       
#> [175] "EBr6"              "IPS(R)"            "EPA(L)"           
#> [178] "AL-VC3l(R)"        "NO3(R)"            "NO3(L)"           
#> [181] "AL-DM6(L)"         "FBl6"              "AOT(R)"           
#> [184] "PB"                "AL-DA2(L)"         "AVLP(R)"          
#> [187] "AB(R)"             "AL-DA4l(R)"        "PENP"             
#> [190] "PB(L9)"            "AL-DA4m(L)"        "ATL(R)"           
#> [193] "ME(R)"             "a'1(R)"            "AL-V(R)"          
#> [196] "AL-DL4(R)"         "GF(R)"             "LAL(-GA)(R)"      
#> [199] "AL-VM7v(L)"        "PB(R9)"            "bL(R)"            
#> [202] "MB(R)"             "AL-VA1v(R)"        "IB"               
#> [205] "g2(R)"             "AL-DC4(L)"         "g5(R)"            
#> [208] "SNP(L)"            "AL-VL1(R)"         "LAL(L)"           
#> [211] "AL-DM5(L)"         "AL-VA6(L)"         "ROB(R)"           
#> [214] "ICL(L)"            "PB(L4)"            "AL-DC1(R)"        
#> [217] "AL-DM5(R)"         "PRW"               "PB(R7)"           
#> [220] "FBl2"              "FBl4"              "AL-DP1l(R)"       
#> [223] "AL-VL2p(R)"        "CX"                "GA(R)"            
#> [226] "VES(R)"            "aL(R)"             "SPS(L)"           
#> [229] "PB(R5)"            "FLA(R)"            "AL-DA3(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PLP(R)"            "LX(L)"             "MB(R)"            
#>   [4] "LAL(-GA)(R)"       "g1(R)"             "FBl3"             
#>   [7] "PB(R4)"            "PB(L4)"            "BU(L)"            
#>  [10] "bL(L)"             "FBl9"              "CAN(R)"           
#>  [13] "AOT(R)"            "RUB(R)"            "b1(R)"            
#>  [16] "LAL(R)"            "GC"                "FB"               
#>  [19] "LX(R)"             "ME(R)"             "EB"               
#>  [22] "gL(R)"             "OL(R)"             "PB(L3)"           
#>  [25] "GOR(L)"            "EBr2r4"            "ROB(R)"           
#>  [28] "NO2(R)"            "aL(L)"             "EPA(R)"           
#>  [31] "PB(R1)"            "CRE(L)"            "LO(R)"            
#>  [34] "AVLP(R)"           "PB(R8)"            "a'1(R)"           
#>  [37] "PB(L7)"            "EBr3pw"            "NO(L)"            
#>  [40] "hemibrain"         "EBr5"              "dACA(R)"          
#>  [43] "FLA(R)"            "EBr3d"             "PB(L6)"           
#>  [46] "a'L(L)"            "POC"               "SCL(L)"           
#>  [49] "a1(R)"             "a'L(R)"            "SAD(-AMMC)"       
#>  [52] "IPS(R)"            "NO3(R)"            "NO1(R)"           
#>  [55] "ICL(L)"            "PED(R)"            "b2(R)"            
#>  [58] "PB(L1)"            "AME(R)"            "GNG"              
#>  [61] "CRE(-ROB,-RUB)(R)" "PB(R6)"            "LAL(L)"           
#>  [64] "EBr3am"            "b'2(R)"            "SAD"              
#>  [67] "NO(R)"             "PB(R3)"            "VLNP(R)"          
#>  [70] "SIP(L)"            "SIP(R)"            "FBl4"             
#>  [73] "ATL(L)"            "mALT(R)"           "PB(R9)"           
#>  [76] "FBl7"              "ICL(R)"            "AB(R)"            
#>  [79] "INP"               "MB(L)"             "EBr1"             
#>  [82] "g5(R)"             "PRW"               "b'L(L)"           
#>  [85] "gL(L)"             "PB(R2)"            "AL(R)"            
#>  [88] "PB(L2)"            "NO2(L)"            "GOR(R)"           
#>  [91] "PB(R5)"            "mALT(L)"           "g2(R)"            
#>  [94] "SCL(R)"            "SPS(L)"            "a'2(R)"           
#>  [97] "NO"                "SNP(L)"            "FBl2"             
#> [100] "LH(R)"             "a3(R)"             "b'1(R)"           
#> [103] "GA(R)"             "VES(R)"            "PVLP(R)"          
#> [106] "CA(R)"             "PB(L5)"            "CA(L)"            
#> [109] "WED(R)"            "PB"                "VMNP"             
#> [112] "g4(R)"             "SLP(R)"            "SMP(L)"           
#> [115] "FBl6"              "ATL(R)"            "a2(R)"            
#> [118] "PB(R7)"            "LOP(R)"            "PB(L8)"           
#> [121] "FBl5"              "FBl8"              "g3(R)"            
#> [124] "SPS(R)"            "PB(L9)"            "VES(L)"           
#> [127] "AMMC"              "GF(R)"             "aL(R)"            
#> [130] "NO1(L)"            "MB(+ACA)(R)"       "vACA(R)"          
#> [133] "CX"                "CRE(R)"            "EBr6"             
#> [136] "SNP(R)"            "BU(R)"             "SMP(R)"           
#> [139] "a'3(R)"            "PENP"              "AOTU(R)"          
#> [142] "lACA(R)"           "b'L(R)"            "RUB(L)"           
#> [145] "bL(R)"             "AL(L)"             "NO3(L)"           
#> [148] "FBl1"              "EPA(L)"            "AB(L)"            
#> [151] "IB"               
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DM5(L)"         "AL-DC4(R)"         "AL-VM1(R)"        
#>   [4] "AL-VP1l(R)"        "a'3(R)"            "MB(L)"            
#>   [7] "ICL(L)"            "AL-VA5(R)"         "SAD"              
#>  [10] "AL-DC3(R)"         "POC"               "VES(R)"           
#>  [13] "IB"                "CRE(R)"            "AL-DL4(R)"        
#>  [16] "FB"                "ROB(R)"            "AL-DA4m(R)"       
#>  [19] "AL-D(R)"           "AOT(R)"            "bL(R)"            
#>  [22] "hemibrain"         "EPA(R)"            "FBl8"             
#>  [25] "VES(L)"            "mALT(L)"           "AL-VP2(R)"        
#>  [28] "PB(L2)"            "FBl9"              "PB(R4)"           
#>  [31] "GOR(R)"            "b'L(L)"            "AL-DC1(R)"        
#>  [34] "AL-V(R)"           "SPS(R)"            "AL-VM7d(L)"       
#>  [37] "RUB(L)"            "PB(R6)"            "mALT(R)"          
#>  [40] "SPS(L)"            "SLP(R)"            "PB(R8)"           
#>  [43] "FBl2"              "NO(L)"             "b2(R)"            
#>  [46] "BU(L)"             "PB(R1)"            "gL(R)"            
#>  [49] "OL(R)"             "AL-DC2(R)"         "AL-VL2a(R)"       
#>  [52] "AL-VP3(R)"         "GOR(L)"            "LX(R)"            
#>  [55] "AL-DM6(L)"         "INP"               "AL-VP1d(R)"       
#>  [58] "vACA(R)"           "AL-DA2(L)"         "AL-DM5(R)"        
#>  [61] "AL-VA4(R)"         "EBr2r4"            "PB(R9)"           
#>  [64] "bL(L)"             "g5(R)"             "FBl5"             
#>  [67] "CA(R)"             "AVLP(R)"           "AL(R)"            
#>  [70] "AL-VA7l(R)"        "AL-DM1(R)"         "CRE(-ROB,-RUB)(R)"
#>  [73] "dACA(R)"           "AL-DP1m(L)"        "RUB(R)"           
#>  [76] "AL-DL2v(R)"        "AL-VC2(R)"         "g2(R)"            
#>  [79] "PB(L4)"            "AL-VA1d(R)"        "AL-DM1(L)"        
#>  [82] "FBl4"              "b1(R)"             "EPA(L)"           
#>  [85] "AL-DA1(R)"         "SMP(R)"            "LAL(L)"           
#>  [88] "PB(L9)"            "LAL(R)"            "AL-VL1(R)"        
#>  [91] "AOTU(R)"           "PB"                "EB"               
#>  [94] "WED(R)"            "PLP(R)"            "a'L(R)"           
#>  [97] "AL-DM4(L)"         "NO"                "PENP"             
#> [100] "AL-D(L)"           "a'1(R)"            "AL-VA6(R)"        
#> [103] "AL-VL2p(R)"        "AL-DL1(R)"         "a'2(R)"           
#> [106] "SAD(-AMMC)"        "IPS(R)"            "AME(R)"           
#> [109] "AL-VC5(R)"         "gL(L)"             "AL-DM4(R)"        
#> [112] "AL-VA7m(R)"        "NO1(L)"            "NO(R)"            
#> [115] "PB(R2)"            "AB(R)"             "AL-VM3(R)"        
#> [118] "g3(R)"             "ME(R)"             "PB(R3)"           
#> [121] "ICL(R)"            "AL-DM3(R)"         "PRW"              
#> [124] "FLA(R)"            "FB-column3"        "LH(R)"            
#> [127] "AL-VP1m(R)"        "EBr3am"            "a2(R)"            
#> [130] "AL-DM3(L)"         "b'L(R)"            "SNP(R)"           
#> [133] "AL-VA6(L)"         "SCL(R)"            "VMNP"             
#> [136] "CAN(R)"            "AL-DM2(R)"         "aL(R)"            
#> [139] "EBr6"              "GNG"               "NO1(R)"           
#> [142] "AL-DA3(L)"         "AL-DA4m(L)"        "AL-DC4(L)"        
#> [145] "AL-DA2(R)"         "AL-VM7v(R)"        "lACA(R)"          
#> [148] "LAL(-GA)(R)"       "PB(L1)"            "AL-DC1(L)"        
#> [151] "AL-DA4l(R)"        "AL-VA3(R)"         "AL-VM7d(R)"       
#> [154] "PB(R5)"            "CX"                "PB(L8)"           
#> [157] "AL-DM2(L)"         "NO3(L)"            "NO2(R)"           
#> [160] "AL-VA2(R)"         "SMP(L)"            "AL-DC2(L)"        
#> [163] "PVLP(R)"           "AL-VA1v(R)"        "CA(L)"            
#> [166] "AL(L)"             "SNP(L)"            "AL-DP1m(R)"       
#> [169] "ATL(L)"            "GF(R)"             "BU(R)"            
#> [172] "AL-DL4(L)"         "CRE(L)"            "AL-DL5(R)"        
#> [175] "GC"                "MB(+ACA)(R)"       "FBl3"             
#> [178] "AL-DP1l(R)"        "SCL(L)"            "SIP(R)"           
#> [181] "PB(L6)"            "AL-VC3l(R)"        "AL-VM5v(R)"       
#> [184] "PB(L7)"            "PB(R7)"            "NO3(R)"           
#> [187] "EBr3pw"            "LX(L)"             "PED(R)"           
#> [190] "GA(R)"             "FBl7"              "AL-VM5d(R)"       
#> [193] "SIP(L)"            "LOP(R)"            "NO2(L)"           
#> [196] "g1(R)"             "a'L(L)"            "AL-DL3(R)"        
#> [199] "b'1(R)"            "ATL(R)"            "aL(L)"            
#> [202] "AL-VP5(R)"         "AL-VP4(R)"         "a3(R)"            
#> [205] "AL-VC1(R)"         "VLNP(R)"           "EBr5"             
#> [208] "AL-VC4(R)"         "PB(L3)"            "FBl1"             
#> [211] "AL-DA3(R)"         "AL-VM7v(L)"        "MB(R)"            
#> [214] "AL-VM2(R)"         "AB(L)"             "CRE(-RUB)(L)"     
#> [217] "AL-VM4(R)"         "PB(L5)"            "EBr1"             
#> [220] "EBr3d"             "g4(R)"             "FBl6"             
#> [223] "AMMC"              "a1(R)"             "AL-DL2d(R)"       
#> [226] "b'2(R)"            "AL-DL5(L)"         "LO(R)"            
#> [229] "AL-DM6(R)"         "AL-VC3m(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "SLP(R)"            "AL-DL2d(R)"        "AL-DM6(R)"        
#>   [4] "AL-DP1l(R)"        "AL-VA4(R)"         "AL-VP1l(R)"       
#>   [7] "POC"               "PENP"              "BU(L)"            
#>  [10] "a'L(R)"            "AL-VM7d(L)"        "LH(R)"            
#>  [13] "NO1(L)"            "b'L(R)"            "AL-DM6(L)"        
#>  [16] "NO3(L)"            "GOR(R)"            "SCL(L)"           
#>  [19] "a'L(L)"            "a1(R)"             "PED(R)"           
#>  [22] "gL(L)"             "AL-DA3(L)"         "AL-DC4(L)"        
#>  [25] "PB(L7)"            "FBl1"              "EBr5"             
#>  [28] "PB(R1)"            "AL-DL4(R)"         "AL-VM7d(R)"       
#>  [31] "a2(R)"             "PB(R7)"            "VMNP"             
#>  [34] "NO(R)"             "ATL(L)"            "aL(L)"            
#>  [37] "LAL(R)"            "AL(R)"             "AL-DL1(R)"        
#>  [40] "AL-DL3(R)"         "AL-VA2(R)"         "AL-VP1m(R)"       
#>  [43] "hemibrain"         "LAL(L)"            "LO(R)"            
#>  [46] "AL-VP2(R)"         "g2(R)"             "AL-VM7v(L)"       
#>  [49] "AL-VC2(R)"         "a'3(R)"            "b'2(R)"           
#>  [52] "NO3(R)"            "SCL(R)"            "EBr2r4"           
#>  [55] "EPA(R)"            "AL-DC1(L)"         "AL-DM2(R)"        
#>  [58] "GOR(L)"            "FLA(R)"            "AL-DM5(L)"        
#>  [61] "NO2(R)"            "AL-DA4l(R)"        "AL-VA3(R)"        
#>  [64] "CA(R)"             "AL-VA6(R)"         "PRW"              
#>  [67] "ATL(R)"            "PB(R3)"            "AL-DA4m(R)"       
#>  [70] "WED(R)"            "AL-DM3(L)"         "g5(R)"            
#>  [73] "FBl7"              "VES(L)"            "AL-DA1(R)"        
#>  [76] "AL-D(R)"           "AL-DM3(R)"         "mALT(R)"          
#>  [79] "ROB(R)"            "PB(L6)"            "bL(L)"            
#>  [82] "SAD(-AMMC)"        "LX(L)"             "MB(R)"            
#>  [85] "VES(R)"            "PVLP(R)"           "FBl2"             
#>  [88] "AME(R)"            "AL-VP4(R)"         "PB(R4)"           
#>  [91] "AL-VC4(R)"         "AL-DM2(L)"         "PLP(R)"           
#>  [94] "EBr3d"             "AL-DM5(R)"         "LOP(R)"           
#>  [97] "SMP(L)"            "FB-column3"        "AL-DA2(R)"        
#> [100] "AL-DC4(R)"         "SPS(R)"            "SPS(L)"           
#> [103] "AL-DA2(L)"         "AL-DM4(L)"         "FBl9"             
#> [106] "AL-VA7l(R)"        "AL-VM3(R)"         "MB(+ACA)(R)"      
#> [109] "AL-VL2p(R)"        "AL-VM5d(R)"        "RUB(R)"           
#> [112] "AL-DC3(R)"         "AL-VM5v(R)"        "AL-VP5(R)"        
#> [115] "AL-DM1(L)"         "OL(R)"             "AL-DA3(R)"        
#> [118] "AL-DC1(R)"         "AL-VA1v(R)"        "AL-V(R)"          
#> [121] "CX"                "AL-VC3m(R)"        "EBr6"             
#> [124] "ME(R)"             "AL-VA7m(R)"        "AL-VM1(R)"        
#> [127] "SIP(R)"            "GC"                "EB"               
#> [130] "AVLP(R)"           "PB(L8)"            "SNP(L)"           
#> [133] "AB(L)"             "CRE(R)"            "vACA(R)"          
#> [136] "AL(L)"             "AL-D(L)"           "CRE(-RUB)(L)"     
#> [139] "bL(R)"             "IB"                "FBl5"             
#> [142] "AL-VP1d(R)"        "PB(R2)"            "SAD"              
#> [145] "dACA(R)"           "AL-DL4(L)"         "PB(L3)"           
#> [148] "lACA(R)"           "NO"                "a3(R)"            
#> [151] "AL-DC2(L)"         "CRE(-ROB,-RUB)(R)" "b1(R)"            
#> [154] "PB(L2)"            "gL(R)"             "PB(L9)"           
#> [157] "PB(R9)"            "PB(L4)"            "AL-VM7v(R)"       
#> [160] "AOTU(R)"           "BU(R)"             "AL-DA4m(L)"       
#> [163] "CAN(R)"            "AL-VM2(R)"         "VLNP(R)"          
#> [166] "EBr3am"            "SIP(L)"            "PB(L1)"           
#> [169] "AL-DM1(R)"         "PB"                "AL-DL2v(R)"       
#> [172] "AOT(R)"            "NO2(L)"            "AB(R)"            
#> [175] "SMP(R)"            "AL-VA6(L)"         "EPA(L)"           
#> [178] "g3(R)"             "FB"                "SNP(R)"           
#> [181] "FBl3"              "AL-VA5(R)"         "g1(R)"            
#> [184] "PB(R8)"            "INP"               "AL-DP1m(R)"       
#> [187] "AL-VA1d(R)"        "AMMC"              "EBr1"             
#> [190] "NO1(R)"            "AL-DP1m(L)"        "b'L(L)"           
#> [193] "NO(L)"             "AL-VC1(R)"         "GF(R)"            
#> [196] "LAL(-GA)(R)"       "MB(L)"             "LX(R)"            
#> [199] "FBl8"              "mALT(L)"           "GA(R)"            
#> [202] "g4(R)"             "EBr3pw"            "IPS(R)"           
#> [205] "CA(L)"             "aL(R)"             "CRE(L)"           
#> [208] "FBl4"              "ICL(R)"            "GNG"              
#> [211] "AL-VL1(R)"         "PB(R6)"            "a'1(R)"           
#> [214] "FBl6"              "AL-DL5(R)"         "AL-VC3l(R)"       
#> [217] "PB(L5)"            "AL-DL5(L)"         "AL-DC2(R)"        
#> [220] "AL-VC5(R)"         "AL-VL2a(R)"        "AL-VM4(R)"        
#> [223] "PB(R5)"            "AL-VP3(R)"         "b2(R)"            
#> [226] "a'2(R)"            "AL-DM4(R)"         "b'1(R)"           
#> [229] "ICL(L)"            "RUB(L)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method   path                                        
#> [1,] "POST"   "/api/v:ver/npexplorer/distribution"        
#> [2,] "POST"   "/api/v:ver/npexplorer/completeness"        
#> [3,] "POST"   "/api/npexplorer/roiconnectivity"           
#> [4,] "REPORT" "/api/*"                                    
#> [5,] "GET"    "/api/v:ver/raw/keyvalue/key/:instance/:key"
#> [6,] "POST"   "/api/v:ver/npexplorer/findneurons"         
#>      name                                                                                           
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIConnectivity-fm"
#> [4,] "github.com/labstack/echo.glob..func1"                                                         
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }