Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "SIP(R)"            "gL(R)"             "g5(R)"            
#>   [4] "CA(R)"             "PB(R3)"            "RUB(L)"           
#>   [7] "GF(R)"             "CRE(R)"            "PB(L8)"           
#>  [10] "RUB(R)"            "hemibrain"         "PB(L7)"           
#>  [13] "LX(L)"             "NO(L)"             "VES(R)"           
#>  [16] "LH(R)"             "FBl9"              "b'L(L)"           
#>  [19] "AB(L)"             "NO1(R)"            "FBl8"             
#>  [22] "bL(L)"             "EBr3pw"            "g2(R)"            
#>  [25] "PB(R9)"            "MB(R)"             "SCL(L)"           
#>  [28] "PB(L2)"            "mALT(L)"           "PENP"             
#>  [31] "aL(L)"             "FLA(R)"            "AL(L)"            
#>  [34] "MB(+ACA)(R)"       "ATL(R)"            "LAL(R)"           
#>  [37] "CA(L)"             "SMP(L)"            "LAL(-GA)(R)"      
#>  [40] "SNP(R)"            "SLP(R)"            "EBr2r4"           
#>  [43] "a'1(R)"            "VLNP(R)"           "SMP(R)"           
#>  [46] "PRW"               "aL(R)"             "AVLP(R)"          
#>  [49] "PB(L4)"            "BU(L)"             "VMNP"             
#>  [52] "GC"                "b'L(R)"            "EB"               
#>  [55] "g3(R)"             "SIP(L)"            "NO"               
#>  [58] "EPA(L)"            "VES(L)"            "GOR(L)"           
#>  [61] "EBr3d"             "b2(R)"             "PB(L9)"           
#>  [64] "LOP(R)"            "CRE(-ROB,-RUB)(R)" "GA(R)"            
#>  [67] "NO3(R)"            "ATL(L)"            "g4(R)"            
#>  [70] "BU(R)"             "PB(R8)"            "FBl3"             
#>  [73] "a3(R)"             "SCL(R)"            "EBr5"             
#>  [76] "LX(R)"             "PB(R1)"            "PB(L3)"           
#>  [79] "g1(R)"             "ROB(R)"            "FBl6"             
#>  [82] "IB"                "EBr3am"            "a1(R)"            
#>  [85] "PB(R6)"            "WED(R)"            "EBr6"             
#>  [88] "PB(L1)"            "CRE(L)"            "LO(R)"            
#>  [91] "ICL(R)"            "AOT(R)"            "lACA(R)"          
#>  [94] "mALT(R)"           "FBl4"              "AB(R)"            
#>  [97] "AL(R)"             "vACA(R)"           "SAD"              
#> [100] "PVLP(R)"           "MB(L)"             "ME(R)"            
#> [103] "a'L(R)"            "PB(L6)"            "b1(R)"            
#> [106] "PB(R4)"            "AOTU(R)"           "SPS(R)"           
#> [109] "NO3(L)"            "FBl1"              "NO2(L)"           
#> [112] "PB"                "a'2(R)"            "POC"              
#> [115] "NO1(L)"            "SPS(L)"            "IPS(R)"           
#> [118] "INP"               "PB(R5)"            "CX"               
#> [121] "ICL(L)"            "FBl5"              "dACA(R)"          
#> [124] "SAD(-AMMC)"        "b'1(R)"            "b'2(R)"           
#> [127] "FB"                "LAL(L)"            "PLP(R)"           
#> [130] "FBl7"              "PB(R2)"            "GOR(R)"           
#> [133] "a'L(L)"            "AMMC"              "PB(R7)"           
#> [136] "gL(L)"             "NO2(R)"            "a'3(R)"           
#> [139] "EPA(R)"            "SNP(L)"            "NO(R)"            
#> [142] "CAN(R)"            "GNG"               "PB(L5)"           
#> [145] "a2(R)"             "EBr1"              "AME(R)"           
#> [148] "bL(R)"             "OL(R)"             "PED(R)"           
#> [151] "FBl2"             
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                               
#> [1,] "PUT"     "/api"                             
#> [2,] "GET"     "/api/v:ver/custom/custom"         
#> [3,] "POST"    "/api/npexplorer/findneurons"      
#> [4,] "POST"    "/api/npexplorer/simpleconnections"
#> [5,] "GET"     "/api/npexplorer/completeness"     
#> [6,] "CONNECT" "/api"                             
#>      name                                                                                             
#> [1,] "github.com/labstack/echo.glob..func1"                                                           
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getFindNeurons-fm"      
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCompleteness-fm"     
#> [6,] "github.com/labstack/echo.glob..func1"                                                           

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }