Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "LO"  "LOP" "ME" 
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "gL(L)"             "AL-VC1(R)"         "AOT(R)"           
#>   [4] "AMMC"              "NO1(L)"            "CRE(R)"           
#>   [7] "gL(R)"             "AL-DC1(R)"         "AL-VM5v(R)"       
#>  [10] "EBr2r4"            "CRE(L)"            "FBl6"             
#>  [13] "PB(R3)"            "AL-DL4(R)"         "AL-VP3(R)"        
#>  [16] "AL-VP2(R)"         "PENP"              "hemibrain"        
#>  [19] "FBl2"              "LAL(R)"            "EBr3am"           
#>  [22] "WED(R)"            "b'L(R)"            "ROB(R)"           
#>  [25] "aL(L)"             "AL-D(R)"           "SPS(R)"           
#>  [28] "EBr6"              "FBl9"              "NO1(R)"           
#>  [31] "PLP(R)"            "PB(L7)"            "AL-VA4(R)"        
#>  [34] "AL-VM7v(L)"        "NO(R)"             "ICL(L)"           
#>  [37] "AL-DL2d(R)"        "AL-DM2(R)"         "AL-VM5d(R)"       
#>  [40] "LAL(-GA)(R)"       "AL-VP5(R)"         "LOP(R)"           
#>  [43] "NO2(R)"            "AL-VM7v(R)"        "RUB(R)"           
#>  [46] "AL-VC3l(R)"        "lACA(R)"           "GOR(R)"           
#>  [49] "a3(R)"             "AL-VA7l(R)"        "AL-VM4(R)"        
#>  [52] "g2(R)"             "BU(R)"             "AL-DC1(L)"        
#>  [55] "bL(R)"             "dACA(R)"           "AL-DM1(L)"        
#>  [58] "AL-VM7d(L)"        "PB(R4)"            "AL-DA4l(R)"       
#>  [61] "PB(R1)"            "FB-column3"        "AB(R)"            
#>  [64] "EPA(L)"            "AL-VM7d(R)"        "OL(R)"            
#>  [67] "LAL(L)"            "EBr3pw"            "b2(R)"            
#>  [70] "b'2(R)"            "GNG"               "SPS(L)"           
#>  [73] "SNP(R)"            "PB(L2)"            "PB(L3)"           
#>  [76] "AL-DL4(L)"         "SCL(L)"            "AL-VA6(L)"        
#>  [79] "PB(L4)"            "AL-DL2v(R)"        "mALT(R)"          
#>  [82] "LO(R)"             "AME(R)"            "INP"              
#>  [85] "GC"                "g5(R)"             "MB(+ACA)(R)"      
#>  [88] "AL-DL1(R)"         "SIP(L)"            "g1(R)"            
#>  [91] "BU(L)"             "FBl3"              "IPS(R)"           
#>  [94] "AL-DA3(R)"         "AL-DA4m(R)"        "AL-DC4(R)"        
#>  [97] "AL-D(L)"           "AL-VA5(R)"         "AL-VC2(R)"        
#> [100] "AL-VL2p(R)"        "mALT(L)"           "SAD"              
#> [103] "LX(R)"             "g4(R)"             "FB"               
#> [106] "FBl5"              "SLP(R)"            "MB(R)"            
#> [109] "AL-VL2a(R)"        "SMP(R)"            "GA(R)"            
#> [112] "VMNP"              "PB(L9)"            "FBl7"             
#> [115] "AL-DM5(R)"         "AL-VA2(R)"         "AL-VP4(R)"        
#> [118] "VLNP(R)"           "SNP(L)"            "PB(R2)"           
#> [121] "AL-DP1l(R)"        "AL-VC5(R)"         "RUB(L)"           
#> [124] "AL-DL5(R)"         "AL-VA6(R)"         "AB(L)"            
#> [127] "vACA(R)"           "NO3(R)"            "PB(L1)"           
#> [130] "AL-DP1m(L)"        "NO3(L)"            "PVLP(R)"          
#> [133] "FLA(R)"            "AL-DA4m(L)"        "AL-DL5(L)"        
#> [136] "AL-DM2(L)"         "CA(R)"             "CRE(-ROB,-RUB)(R)"
#> [139] "CX"                "PB(R6)"            "AL-DM4(L)"        
#> [142] "AL-DM4(R)"         "a'3(R)"            "IB112"            
#> [145] "CAN(R)"            "b'L(L)"            "b'1(R)"           
#> [148] "EBr5"              "EPA(R)"            "AL-DC2(L)"        
#> [151] "CRE(-RUB)(L)"      "FBl4"              "PB(R8)"           
#> [154] "PB"                "bL(L)"             "AVLP(R)"          
#> [157] "VES(L)"            "AL-VA1v(R)"        "AL-VP1m(R)"       
#> [160] "SMP(L)"            "g3(R)"             "AOTU(R)"          
#> [163] "AL-DP1m(R)"        "PB(R9)"            "LX(L)"            
#> [166] "PRW"               "GOR(L)"            "AL-DA3(L)"        
#> [169] "PB(L6)"            "a'1(R)"            "AL-DC4(L)"        
#> [172] "FBl1"              "b1(R)"             "NO(L)"            
#> [175] "AL-VC4(R)"         "PB(R5)"            "a'L(R)"           
#> [178] "FBl8"              "ATL(R)"            "AL-DC3(R)"        
#> [181] "AL-DM1(R)"         "aL(R)"             "SCL(R)"           
#> [184] "AL-VA3(R)"         "AL-VA7m(R)"        "ATL(L)"           
#> [187] "AL-DM3(L)"         "AL-DM5(L)"         "AL-DC2(R)"        
#> [190] "NO2(L)"            "PED(R)"            "AL-DM6(L)"        
#> [193] "AL-VA1d(R)"        "AL-VP1d(R)"        "PB(L5)"           
#> [196] "NO"                "GF(R)"             "SIP(R)"           
#> [199] "SAD(-AMMC)"        "AL(R)"             "AL-DA2(R)"        
#> [202] "AL-DM3(R)"         "AL-VM1(R)"         "AL-VM3(R)"        
#> [205] "AL-V(R)"           "a2(R)"             "PB(R7)"           
#> [208] "ME(R)"             "ICL(R)"            "EBr1"             
#> [211] "AL(L)"             "LH(R)"             "a'L(L)"           
#> [214] "AL-VC3m(R)"        "AL-VM2(R)"         "CA(L)"            
#> [217] "MB(L)"             "EB"                "AL-DA1(R)"        
#> [220] "AL-VL1(R)"         "PB(L8)"            "EBr3d"            
#> [223] "AL-DM6(R)"         "a'2(R)"            "AL-DA2(L)"        
#> [226] "AL-DL3(R)"         "IB"                "a1(R)"            
#> [229] "VES(R)"            "AL-VP1l(R)"        "POC"              
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "GF(R)"             "SNP(R)"            "ICL(L)"           
#>   [4] "CX"                "g3(R)"             "GOR(L)"           
#>   [7] "gL(R)"             "MB(L)"             "PB(L6)"           
#>  [10] "LAL(R)"            "PB(L4)"            "OL(R)"            
#>  [13] "AOT(R)"            "PB(L1)"            "EBr2r4"           
#>  [16] "SAD(-AMMC)"        "PB(L9)"            "FBl4"             
#>  [19] "FBl6"              "AL(R)"             "GC"               
#>  [22] "SCL(R)"            "a'2(R)"            "FBl8"             
#>  [25] "MB(R)"             "CRE(-ROB,-RUB)(R)" "CA(L)"            
#>  [28] "PB(L8)"            "EPA(R)"            "AME(R)"           
#>  [31] "AB(L)"             "LAL(-GA)(R)"       "VES(R)"           
#>  [34] "LX(L)"             "LX(R)"             "BU(R)"            
#>  [37] "FLA(R)"            "PB"                "VMNP"             
#>  [40] "NO3(L)"            "PENP"              "aL(R)"            
#>  [43] "EBr3d"             "EBr3am"            "LOP(R)"           
#>  [46] "PB(L7)"            "MB(+ACA)(R)"       "ATL(R)"           
#>  [49] "CAN(R)"            "aL(L)"             "RUB(L)"           
#>  [52] "a'L(L)"            "PB(R7)"            "dACA(R)"          
#>  [55] "NO2(R)"            "SIP(L)"            "PED(R)"           
#>  [58] "ROB(R)"            "g2(R)"             "PB(R8)"           
#>  [61] "ME(R)"             "LO(R)"             "NO(L)"            
#>  [64] "SPS(R)"            "mALT(R)"           "WED(R)"           
#>  [67] "POC"               "IB"                "vACA(R)"          
#>  [70] "EBr5"              "BU(L)"             "FBl2"             
#>  [73] "SMP(L)"            "g4(R)"             "a'3(R)"           
#>  [76] "EBr1"              "mALT(L)"           "AMMC"             
#>  [79] "NO1(L)"            "NO1(R)"            "PLP(R)"           
#>  [82] "AVLP(R)"           "PB(L5)"            "PB(L3)"           
#>  [85] "PRW"               "g1(R)"             "bL(R)"            
#>  [88] "GOR(R)"            "RUB(R)"            "a'L(R)"           
#>  [91] "SNP(L)"            "CA(R)"             "SPS(L)"           
#>  [94] "CRE(L)"            "SMP(R)"            "PB(R6)"           
#>  [97] "b'L(R)"            "EBr3pw"            "FB"               
#> [100] "PB(R1)"            "EPA(L)"            "FBl7"             
#> [103] "PB(R4)"            "ATL(L)"            "NO2(L)"           
#> [106] "b'2(R)"            "SAD"               "PB(R9)"           
#> [109] "FBl1"              "FBl5"              "INP"              
#> [112] "SLP(R)"            "SCL(L)"            "AOTU(R)"          
#> [115] "SIP(R)"            "a'1(R)"            "FBl9"             
#> [118] "LAL(L)"            "g5(R)"             "FBl3"             
#> [121] "ICL(R)"            "AB(R)"             "VES(L)"           
#> [124] "b'1(R)"            "NO3(R)"            "NO"               
#> [127] "PB(R2)"            "EB"                "GA(R)"            
#> [130] "lACA(R)"           "EBr6"              "PB(R5)"           
#> [133] "b2(R)"             "NO(R)"             "GNG"              
#> [136] "a3(R)"             "a1(R)"             "hemibrain"        
#> [139] "PVLP(R)"           "b1(R)"             "AL(L)"            
#> [142] "VLNP(R)"           "PB(L2)"            "IPS(R)"           
#> [145] "gL(L)"             "LH(R)"             "PB(R3)"           
#> [148] "b'L(L)"            "a2(R)"             "CRE(R)"           
#> [151] "bL(L)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "VLNP(R)"           "FBl1"              "g4(R)"            
#>   [4] "PB(L3)"            "AL-DC4(L)"         "PVLP(R)"          
#>   [7] "GNG"               "AL-DM4(R)"         "MB(L)"            
#>  [10] "SIP(R)"            "PB(R1)"            "GOR(R)"           
#>  [13] "lACA(R)"           "a'3(R)"            "SNP(R)"           
#>  [16] "AL-DM4(L)"         "PB(L1)"            "PB(L8)"           
#>  [19] "AL-DC2(L)"         "AL-DL4(L)"         "g3(R)"            
#>  [22] "NO1(R)"            "b1(R)"             "AL-VA5(R)"        
#>  [25] "AL-VC1(R)"         "PRW"               "PB(R2)"           
#>  [28] "LX(L)"             "AL-VM5v(R)"        "CRE(L)"           
#>  [31] "PB(R4)"            "AVLP(R)"           "AL-VA7m(R)"       
#>  [34] "bL(R)"             "PB"                "NO(R)"            
#>  [37] "SNP(L)"            "LH(R)"             "NO2(L)"           
#>  [40] "GA(R)"             "CRE(-RUB)(L)"      "AL-DL5(R)"        
#>  [43] "PENP"              "AL-VC3m(R)"        "PB(L5)"           
#>  [46] "ATL(L)"            "ATL(R)"            "AL-DA4m(L)"       
#>  [49] "g2(R)"             "AL-DM5(L)"         "AB(R)"            
#>  [52] "FBl6"              "AL-VM4(R)"         "AL(R)"            
#>  [55] "AL-DM6(R)"         "AL-VA2(R)"         "AL-DL5(L)"        
#>  [58] "EBr2r4"            "AL(L)"             "AL-VM5d(R)"       
#>  [61] "AL-VM7d(R)"        "BU(L)"             "AL-DL3(R)"        
#>  [64] "AL-VM2(R)"         "AL-VP5(R)"         "EPA(R)"           
#>  [67] "a2(R)"             "PB(R8)"            "BU(R)"            
#>  [70] "MB(R)"             "AL-DM2(R)"         "PB(R6)"           
#>  [73] "AL-DM2(L)"         "AL-VM7v(L)"        "AL-VM7v(R)"       
#>  [76] "AMMC"              "gL(L)"             "SCL(L)"           
#>  [79] "SIP(L)"            "VES(L)"            "AL-D(R)"          
#>  [82] "FBl7"              "b2(R)"             "ME(R)"            
#>  [85] "PED(R)"            "LX(R)"             "AL-DL1(R)"        
#>  [88] "AOT(R)"            "CRE(-ROB,-RUB)(R)" "FLA(R)"           
#>  [91] "AL-DC1(L)"         "AL-DM6(L)"         "NO"               
#>  [94] "AL-VA6(R)"         "SPS(L)"            "FBl3"             
#>  [97] "PB(L4)"            "AL-VA4(R)"         "ICL(L)"           
#> [100] "AL-V(R)"           "PB(R9)"            "PB(L6)"           
#> [103] "PB(R3)"            "aL(L)"             "INP"              
#> [106] "AL-DL2d(R)"        "EBr6"              "AL-VA6(L)"        
#> [109] "AL-DA4m(R)"        "AL-VA1d(R)"        "GOR(L)"           
#> [112] "EPA(L)"            "AL-VL2a(R)"        "PLP(R)"           
#> [115] "AOTU(R)"           "VMNP"              "PB(L7)"           
#> [118] "ICL(R)"            "AL-DM1(R)"         "b'2(R)"           
#> [121] "a'2(R)"            "AL-DM1(L)"         "SMP(L)"           
#> [124] "NO3(L)"            "g5(R)"             "FBl8"             
#> [127] "IB"                "LAL(-GA)(R)"       "PB(R7)"           
#> [130] "GC"                "ROB(R)"            "FBl9"             
#> [133] "AL-DP1m(R)"        "PB(R5)"            "hemibrain"        
#> [136] "CRE(R)"            "AL-DP1m(L)"        "AL-DP1l(R)"       
#> [139] "AL-VL2p(R)"        "WED(R)"            "dACA(R)"          
#> [142] "VES(R)"            "CA(R)"             "AL-VP2(R)"        
#> [145] "AL-VP1l(R)"        "GF(R)"             "AL-VP1m(R)"       
#> [148] "AL-DA1(R)"         "AL-DL4(R)"         "AB(L)"            
#> [151] "a1(R)"             "EBr3pw"            "AL-DM3(L)"        
#> [154] "AL-VM3(R)"         "EBr1"              "NO2(R)"           
#> [157] "a'L(L)"            "AL-DM3(R)"         "AL-VC3l(R)"       
#> [160] "AL-VP4(R)"         "AL-VP1d(R)"        "FB-column3"       
#> [163] "EB"                "AL-DA2(R)"         "AL-DC4(R)"        
#> [166] "AL-VA7l(R)"        "vACA(R)"           "mALT(R)"          
#> [169] "a3(R)"             "AL-DA4l(R)"        "gL(R)"            
#> [172] "PB(L9)"            "a'1(R)"            "AL-VC2(R)"        
#> [175] "EBr3am"            "CA(L)"             "NO1(L)"           
#> [178] "AL-VC5(R)"         "a'L(R)"            "AL-VL1(R)"        
#> [181] "b'L(L)"            "PB(L2)"            "SPS(R)"           
#> [184] "AL-DA3(R)"         "AL-D(L)"           "AL-DC2(R)"        
#> [187] "AL-DL2v(R)"        "AL-DM5(R)"         "AL-VA1v(R)"       
#> [190] "AL-VC4(R)"         "LOP(R)"            "EBr5"             
#> [193] "CAN(R)"            "FBl2"              "AME(R)"           
#> [196] "NO(L)"             "LAL(R)"            "g1(R)"            
#> [199] "b'L(R)"            "AL-DC3(R)"         "FB"               
#> [202] "SCL(R)"            "OL(R)"             "SLP(R)"           
#> [205] "AL-DC1(R)"         "POC"               "mALT(L)"          
#> [208] "aL(R)"             "AL-VP3(R)"         "EBr3d"            
#> [211] "bL(L)"             "SAD(-AMMC)"        "LO(R)"            
#> [214] "RUB(L)"            "AL-DA3(L)"         "SAD"              
#> [217] "AL-VM7d(L)"        "MB(+ACA)(R)"       "FBl4"             
#> [220] "AL-VM1(R)"         "SMP(R)"            "LAL(L)"           
#> [223] "AL-VA3(R)"         "IPS(R)"            "AL-DA2(L)"        
#> [226] "FBl5"              "RUB(R)"            "b'1(R)"           
#> [229] "CX"                "NO3(R)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "SIP(L)"            "GC"                "a'L(R)"           
#>   [4] "RUB(L)"            "MB(R)"             "a'2(R)"           
#>   [7] "gL(R)"             "PB(R5)"            "SCL(R)"           
#>  [10] "PB(L7)"            "AOT(R)"            "NO2(L)"           
#>  [13] "PB(L3)"            "PB(L6)"            "AL-VC3m(R)"       
#>  [16] "mALT(R)"           "AL-DL5(L)"         "NO"               
#>  [19] "CRE(L)"            "dACA(R)"           "AL(R)"            
#>  [22] "AL-DL2d(R)"        "AL-DL3(R)"         "AL-VA4(R)"        
#>  [25] "AL-DA3(L)"         "g5(R)"             "AL-DC3(R)"        
#>  [28] "AL-DM1(L)"         "a'1(R)"            "AL-DM1(R)"        
#>  [31] "BU(L)"             "lACA(R)"           "AL-VA1d(R)"       
#>  [34] "AL-VL2a(R)"        "AL-DM3(L)"         "FB-column3"       
#>  [37] "a'3(R)"            "FBl3"              "VES(L)"           
#>  [40] "AL-VM3(R)"         "VLNP(R)"           "PRW"              
#>  [43] "NO1(R)"            "FBl7"              "a2(R)"            
#>  [46] "a3(R)"             "AL-DL5(R)"         "IB"               
#>  [49] "AL-DA4m(L)"        "AL-DM5(R)"         "mALT(L)"          
#>  [52] "CX"                "LO(R)"             "AL-VP5(R)"        
#>  [55] "CA(R)"             "EBr3am"            "SPS(R)"           
#>  [58] "AL(L)"             "SNP(L)"            "a'L(L)"           
#>  [61] "AL-VM5v(R)"        "NO3(L)"            "PB(L5)"           
#>  [64] "EBr3d"             "PVLP(R)"           "PB(R4)"           
#>  [67] "AL-DL1(R)"         "AL-VM2(R)"         "AMMC"             
#>  [70] "SAD"               "WED(R)"            "VMNP"             
#>  [73] "AL-VM7v(L)"        "AL-DA4l(R)"        "AL-VC3l(R)"       
#>  [76] "PB(L2)"            "PB"                "PB(R1)"           
#>  [79] "FBl9"              "NO2(R)"            "AL-VP1d(R)"       
#>  [82] "PB(R6)"            "FB"                "AL-DC2(R)"        
#>  [85] "AL-VM1(R)"         "LAL(L)"            "NO3(R)"           
#>  [88] "EBr1"              "bL(L)"             "NO(L)"            
#>  [91] "AL-VA7m(R)"        "GOR(L)"            "AL-DM6(L)"        
#>  [94] "b1(R)"             "AL-DL2v(R)"        "AL-DM4(L)"        
#>  [97] "FBl8"              "AL-VP1l(R)"        "aL(L)"            
#> [100] "AL-DA1(R)"         "AL-VP1m(R)"        "NO1(L)"           
#> [103] "AL-DC2(L)"         "AL-VC5(R)"         "AL-DC4(L)"        
#> [106] "PB(L9)"            "VES(R)"            "PB(R2)"           
#> [109] "SCL(L)"            "AL-DC1(L)"         "AL-DM3(R)"        
#> [112] "PB(L1)"            "MB(+ACA)(R)"       "SMP(R)"           
#> [115] "aL(R)"             "MB(L)"             "FLA(R)"           
#> [118] "b'L(L)"            "ICL(L)"            "AL-VA1v(R)"       
#> [121] "ME(R)"             "AL-DM2(L)"         "GF(R)"            
#> [124] "AL-DC4(R)"         "AL-VC2(R)"         "AL-VP3(R)"        
#> [127] "hemibrain"         "g4(R)"             "FBl2"             
#> [130] "EPA(L)"            "OL(R)"             "a1(R)"            
#> [133] "EBr2r4"            "AL-VA3(R)"         "AL-V(R)"          
#> [136] "AL-VP2(R)"         "RUB(R)"            "PED(R)"           
#> [139] "LX(R)"             "SLP(R)"            "AL-DL4(R)"        
#> [142] "CA(L)"             "GA(R)"             "LAL(R)"           
#> [145] "AL-DA2(R)"         "AL-DA3(R)"         "PENP"             
#> [148] "EBr6"              "EPA(R)"            "FBl6"             
#> [151] "bL(R)"             "b'L(R)"            "PLP(R)"           
#> [154] "AL-DP1m(L)"        "ROB(R)"            "ICL(R)"           
#> [157] "INP"               "AL-VC1(R)"         "b'2(R)"           
#> [160] "AL-DA2(L)"         "SIP(R)"            "FBl5"             
#> [163] "AL-DP1l(R)"        "AL-DM5(L)"         "FBl1"             
#> [166] "CRE(-ROB,-RUB)(R)" "SPS(L)"            "LH(R)"            
#> [169] "CRE(R)"            "AL-DM4(R)"         "AL-VA6(R)"        
#> [172] "POC"               "AOTU(R)"           "PB(R7)"           
#> [175] "g1(R)"             "CAN(R)"            "PB(L4)"           
#> [178] "AL-DA4m(R)"        "PB(R9)"            "GNG"              
#> [181] "AL-VL2p(R)"        "AL-VM7v(R)"        "b'1(R)"           
#> [184] "LAL(-GA)(R)"       "PB(L8)"            "BU(R)"            
#> [187] "AL-DL4(L)"         "AL-VA5(R)"         "PB(R3)"           
#> [190] "AL-VM4(R)"         "SMP(L)"            "EBr5"             
#> [193] "SAD(-AMMC)"        "AL-D(L)"           "NO(R)"            
#> [196] "ATL(R)"            "GOR(R)"            "AL-DM2(R)"        
#> [199] "AL-VP4(R)"         "ATL(L)"            "gL(L)"            
#> [202] "FBl4"              "AL-DC1(R)"         "vACA(R)"          
#> [205] "AL-VA6(L)"         "AVLP(R)"           "AL-DP1m(R)"       
#> [208] "AL-VC4(R)"         "LOP(R)"            "PB(R8)"           
#> [211] "SNP(R)"            "AL-DM6(R)"         "AL-VA2(R)"        
#> [214] "AB(L)"             "b2(R)"             "AL-VM7d(R)"       
#> [217] "IPS(R)"            "LX(L)"             "AL-VA7l(R)"       
#> [220] "AL-VL1(R)"         "AL-VM5d(R)"        "g3(R)"            
#> [223] "CRE(-RUB)(L)"      "g2(R)"             "EBr3pw"           
#> [226] "AB(R)"             "AME(R)"            "AL-D(R)"          
#> [229] "EB"                "AL-VM7d(L)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method   path                                          
#> [1,] "POST"   "/logout"                                     
#> [2,] "DELETE" "/api/*"                                      
#> [3,] "POST"   "/api/v:ver/raw/cypher/transaction/:id/commit"
#> [4,] "GET"    "/api/v:ver/dbmeta/datasets"                  
#> [5,] "POST"   "/api/v:ver/custom/custom"                    
#> [6,] "GET"    "/api/npexplorer/simpleconnections"           
#>      name                                                                                             
#> [1,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                         
#> [2,] "github.com/labstack/echo/v4.glob..func1"                                                        
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }