Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(L6)"            "LOP(R)"            "AL(L)"            
#>   [4] "OL(R)"             "NO2(R)"            "MB(R)"            
#>   [7] "SCL(L)"            "ATL(L)"            "CA(L)"            
#>  [10] "PB(L8)"            "ME(R)"             "EBr3am"           
#>  [13] "POC"               "AB(L)"             "ICL(L)"           
#>  [16] "PB(L3)"            "a2(R)"             "LX(L)"            
#>  [19] "CAN(R)"            "GOR(R)"            "a3(R)"            
#>  [22] "PB(R7)"            "LX(R)"             "FBl6"             
#>  [25] "FBl8"              "AOT(R)"            "SMP(R)"           
#>  [28] "b'1(R)"            "vACA(R)"           "mALT(R)"          
#>  [31] "PB(L1)"            "NO1(L)"            "FLA(R)"           
#>  [34] "PED(R)"            "NO3(R)"            "aL(L)"            
#>  [37] "FBl9"              "IPS(R)"            "SIP(R)"           
#>  [40] "SNP(L)"            "b'L(L)"            "SCL(R)"           
#>  [43] "PB(R8)"            "SNP(R)"            "NO(R)"            
#>  [46] "LO(R)"             "a'L(L)"            "CRE(R)"           
#>  [49] "EBr5"              "EBr3pw"            "ICL(R)"           
#>  [52] "INP"               "PENP"              "SMP(L)"           
#>  [55] "a'L(R)"            "FBl7"              "NO1(R)"           
#>  [58] "PLP(R)"            "VES(R)"            "FBl1"             
#>  [61] "FBl4"              "AME(R)"            "AL(R)"            
#>  [64] "b'2(R)"            "PB(R9)"            "VMNP"             
#>  [67] "gL(L)"             "FBl2"              "PRW"              
#>  [70] "lACA(R)"           "PB(R1)"            "AB(R)"            
#>  [73] "VLNP(R)"           "SIP(L)"            "BU(R)"            
#>  [76] "EBr2r4"            "PB(L7)"            "PVLP(R)"          
#>  [79] "aL(R)"             "EBr3d"             "hemibrain"        
#>  [82] "WED(R)"            "PB"                "EPA(L)"           
#>  [85] "RUB(L)"            "IB"                "PB(R6)"           
#>  [88] "GA(R)"             "GF(R)"             "PB(R2)"           
#>  [91] "ATL(R)"            "PB(R3)"            "LAL(-GA)(R)"      
#>  [94] "NO"                "GOR(L)"            "a'1(R)"           
#>  [97] "CX"                "GC"                "gL(R)"            
#> [100] "SLP(R)"            "RUB(R)"            "g2(R)"            
#> [103] "SPS(L)"            "CA(R)"             "PB(R5)"           
#> [106] "PB(L2)"            "g3(R)"             "b2(R)"            
#> [109] "g4(R)"             "CRE(L)"            "bL(R)"            
#> [112] "mALT(L)"           "BU(L)"             "dACA(R)"          
#> [115] "NO3(L)"            "FBl5"              "GNG"              
#> [118] "PB(L5)"            "NO2(L)"            "LAL(L)"           
#> [121] "SAD"               "PB(L9)"            "SAD(-AMMC)"       
#> [124] "FBl3"              "LH(R)"             "VES(L)"           
#> [127] "CRE(-ROB,-RUB)(R)" "AOTU(R)"           "SPS(R)"           
#> [130] "MB(+ACA)(R)"       "b1(R)"             "EBr1"             
#> [133] "EBr6"              "NO(L)"             "EPA(R)"           
#> [136] "AVLP(R)"           "a'3(R)"            "b'L(R)"           
#> [139] "ROB(R)"            "EB"                "bL(L)"            
#> [142] "g5(R)"             "LAL(R)"            "AMMC"             
#> [145] "g1(R)"             "MB(L)"             "PB(R4)"           
#> [148] "PB(L4)"            "a1(R)"             "FB"               
#> [151] "a'2(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "FBl1"              "AL-VA1d(R)"        "AL-VA4(R)"        
#>   [4] "AL-VM5v(R)"        "AL-VP1d(R)"        "SAD(-AMMC)"       
#>   [7] "AOTU(R)"           "PED(R)"            "CRE(R)"           
#>  [10] "PB"                "AL-VM1(R)"         "RUB(R)"           
#>  [13] "PENP"              "CRE(L)"            "FBl4"             
#>  [16] "RUB(L)"            "g2(R)"             "WED(R)"           
#>  [19] "AL-VM7v(L)"        "ROB(R)"            "g3(R)"            
#>  [22] "AL-DC1(L)"         "AL-DL4(L)"         "AL-DM1(L)"        
#>  [25] "PLP(R)"            "ATL(R)"            "AL-DL2d(R)"       
#>  [28] "AL-VP5(R)"         "a'3(R)"            "SMP(L)"           
#>  [31] "GOR(R)"            "AL-DA4l(R)"        "GOR(L)"           
#>  [34] "AL-VA7l(R)"        "AL-VM5d(R)"        "ATL(L)"           
#>  [37] "PB(R4)"            "AL-VC1(R)"         "AL-VM2(R)"        
#>  [40] "AL-VM3(R)"         "PB(R7)"            "MB(+ACA)(R)"      
#>  [43] "PB(R3)"            "bL(R)"             "g4(R)"            
#>  [46] "CRE(-ROB,-RUB)(R)" "EBr5"              "AL-DA4m(R)"       
#>  [49] "AL-DL5(L)"         "AL-DM2(L)"         "PB(R2)"           
#>  [52] "g1(R)"             "AL(L)"             "PB(L8)"           
#>  [55] "CAN(R)"            "PB(L4)"            "PRW"              
#>  [58] "FBl3"              "ICL(R)"            "AL-VA5(R)"        
#>  [61] "AL-VM7v(R)"        "SIP(R)"            "LAL(-GA)(R)"      
#>  [64] "PB(L3)"            "GC"                "FBl9"             
#>  [67] "OL(R)"             "AL-VA2(R)"         "AL-V(R)"          
#>  [70] "AL-VP2(R)"         "SAD"               "mALT(R)"          
#>  [73] "AL-VM7d(L)"        "EPA(L)"            "AL-VP4(R)"        
#>  [76] "VLNP(R)"           "POC"               "a1(R)"            
#>  [79] "hemibrain"         "AL-DM4(L)"         "NO(R)"            
#>  [82] "NO3(L)"            "AL-DA1(R)"         "AL-DA3(R)"        
#>  [85] "NO2(L)"            "PB(L1)"            "a'L(L)"           
#>  [88] "vACA(R)"           "AL-VA6(R)"         "CA(R)"            
#>  [91] "gL(R)"             "AL-DM4(R)"         "CA(L)"            
#>  [94] "SCL(R)"            "AL-VL1(R)"         "AL-DC2(R)"        
#>  [97] "bL(L)"             "a2(R)"             "a'2(R)"           
#> [100] "a'L(R)"            "NO(L)"             "AL-DP1l(R)"       
#> [103] "AL-VC2(R)"         "GA(R)"             "aL(L)"            
#> [106] "SLP(R)"            "AL-DL4(R)"         "AL-VC5(R)"        
#> [109] "PB(L2)"            "LAL(L)"            "AL-DL5(R)"        
#> [112] "AL-VL2p(R)"        "LO(R)"             "FBl5"             
#> [115] "AL-DP1m(R)"        "AL-VL2a(R)"        "PB(R5)"           
#> [118] "b'1(R)"            "NO3(R)"            "AOT(R)"           
#> [121] "LOP(R)"            "AL-DC3(R)"         "b'L(L)"           
#> [124] "VES(L)"            "GNG"               "AL-VP1m(R)"       
#> [127] "AME(R)"            "AL-DA2(R)"         "AL-DC1(R)"        
#> [130] "VMNP"              "lACA(R)"           "PB(L9)"           
#> [133] "PB(L7)"            "AMMC"              "SPS(R)"           
#> [136] "EBr3am"            "EBr1"              "FB-column3"       
#> [139] "EPA(R)"            "AB(L)"             "AL-DL3(R)"        
#> [142] "a'1(R)"            "IPS(R)"            "MB(R)"            
#> [145] "AL-DC4(R)"         "AL-DM2(R)"         "GF(R)"            
#> [148] "EBr2r4"            "NO2(R)"            "LX(R)"            
#> [151] "SNP(L)"            "LX(L)"             "AL-VP1l(R)"       
#> [154] "aL(R)"             "gL(L)"             "AL-DC2(L)"        
#> [157] "AL-VA1v(R)"        "SNP(R)"            "BU(R)"            
#> [160] "AL-DA3(L)"         "FBl7"              "b2(R)"            
#> [163] "b'L(R)"            "AL-DA2(L)"         "SMP(R)"           
#> [166] "FB"                "SCL(L)"            "CX"               
#> [169] "AL-DM3(L)"         "AL-DP1m(L)"        "PB(L6)"           
#> [172] "PB(R1)"            "LH(R)"             "AVLP(R)"          
#> [175] "AL(R)"             "AL-DL1(R)"         "BU(L)"            
#> [178] "AL-VC3m(R)"        "AL-DM5(L)"         "PB(R9)"           
#> [181] "FBl2"              "FBl6"              "INP"              
#> [184] "AL-VP3(R)"         "SPS(L)"            "AL-DC4(L)"        
#> [187] "NO"                "AL-VA7m(R)"        "EBr6"             
#> [190] "AL-D(R)"           "AL-VM7d(R)"        "EBr3d"            
#> [193] "AL-VC4(R)"         "AL-DL2v(R)"        "CRE(-RUB)(L)"     
#> [196] "NO1(R)"            "a3(R)"             "AL-DM1(R)"        
#> [199] "PB(R6)"            "LAL(R)"            "AL-DM6(R)"        
#> [202] "AL-VA3(R)"         "EB"                "AL-DM6(L)"        
#> [205] "AL-VM4(R)"         "ME(R)"             "AL-VA6(L)"        
#> [208] "b'2(R)"            "AL-VC3l(R)"        "PVLP(R)"          
#> [211] "dACA(R)"           "PB(L5)"            "MB(L)"            
#> [214] "NO1(L)"            "PB(R8)"            "AL-D(L)"          
#> [217] "AL-DM3(R)"         "SIP(L)"            "g5(R)"            
#> [220] "AL-DA4m(L)"        "ICL(L)"            "FBl8"             
#> [223] "AB(R)"             "AL-DM5(R)"         "mALT(L)"          
#> [226] "FLA(R)"            "VES(R)"            "b1(R)"            
#> [229] "IB"                "EBr3pw"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method   path                               
#> [1,] "GET"    "/api/v:ver/dbmeta/database"       
#> [2,] "POST"   "/api/roimeshes/mesh/:dataset/:roi"
#> [3,] "REPORT" "/api/*"                           
#> [4,] "POST"   "/api/v:ver/raw/cypher/transaction"
#> [5,] "GET"    "/api/*"                           
#> [6,] "POST"   "/api/v:ver/npexplorer/rankedtable"
#>      name                                                                    
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    
#> [2,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"
#> [3,] "github.com/labstack/echo.glob..func1"                                  
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    
#> [5,] "github.com/labstack/echo.glob..func1"                                  
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }