Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $fib19
#> $fib19$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $fib19$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $fib19$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $fib19$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $fib19$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $fib19$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "GOR(L)"            "NO(R)"             "AL-DA1(R)"        
#>   [4] "AL-VA1d(R)"        "AL-DM4(R)"         "IB"               
#>   [7] "CRE(L)"            "FBl2"              "AME(R)"           
#>  [10] "bL(R)"             "AL-DC2(R)"         "AL-VC1(R)"        
#>  [13] "a'L(R)"            "FBl8"              "IPS(R)"           
#>  [16] "b'L(L)"            "AL(R)"             "SCL(L)"           
#>  [19] "FB"                "b'L(R)"            "IB112"            
#>  [22] "FBl5"              "AL-VP1m(R)"        "NO1(R)"           
#>  [25] "gL(L)"             "PB(R1)"            "NO2(L)"           
#>  [28] "AL-DM3(L)"         "MB(+ACA)(R)"       "AL-DM2(R)"        
#>  [31] "PENP"              "AL-DL4(L)"         "AL-VC2(R)"        
#>  [34] "AL-VC4(R)"         "EBr3am"            "AL-D(L)"          
#>  [37] "AL-VC5(R)"         "AL(L)"             "EBr3pw"           
#>  [40] "LX(L)"             "GOR(R)"            "AL-VM5v(R)"       
#>  [43] "g1(R)"             "PB(L9)"            "SLP(R)"           
#>  [46] "AL-VA2(R)"         "CRE(-ROB,-RUB)(R)" "AVLP(R)"          
#>  [49] "a3(R)"             "PED(R)"            "AL-DL5(L)"        
#>  [52] "LO(R)"             "AL-VC3m(R)"        "AL-VL2a(R)"       
#>  [55] "BU(L)"             "AL-VM7d(L)"        "aL(L)"            
#>  [58] "AL-VM7v(R)"        "PB(L1)"            "CA(R)"            
#>  [61] "g2(R)"             "AL-VP2(R)"         "AL-VP1d(R)"       
#>  [64] "AL-VM7d(R)"        "a2(R)"             "NO(L)"            
#>  [67] "AL-D(R)"           "AL-VA1v(R)"        "AL-VM3(R)"        
#>  [70] "NO3(R)"            "MB(L)"             "AL-DL5(R)"        
#>  [73] "g4(R)"             "AL-DA4m(L)"        "CRE(-RUB)(L)"     
#>  [76] "AL-DM1(R)"         "PB(R7)"            "dACA(R)"          
#>  [79] "MB(R)"             "AL-VP5(R)"         "POC"              
#>  [82] "EBr3d"             "ROB(R)"            "PB(L7)"           
#>  [85] "FB-column3"        "CA(L)"             "EBr5"             
#>  [88] "PB(R9)"            "SMP(L)"            "GA(R)"            
#>  [91] "CAN(R)"            "AL-VP1l(R)"        "b'1(R)"           
#>  [94] "NO1(L)"            "EPA(R)"            "PB(R3)"           
#>  [97] "AL-DC4(R)"         "PB(R8)"            "AL-VA3(R)"        
#> [100] "AL-VA7m(R)"        "OL(R)"             "vACA(R)"          
#> [103] "SNP(R)"            "AL-DM5(L)"         "NO"               
#> [106] "AL-VA6(R)"         "a'3(R)"            "LAL(-GA)(R)"      
#> [109] "aL(R)"             "ME(R)"             "BU(R)"            
#> [112] "SIP(R)"            "lACA(R)"           "PVLP(R)"          
#> [115] "AL-VL2p(R)"        "AMMC"              "AL-DM2(L)"        
#> [118] "AL-VM7v(L)"        "AL-DA2(R)"         "AL-V(R)"          
#> [121] "AB(L)"             "PB(R6)"            "AL-DC1(L)"        
#> [124] "g5(R)"             "RUB(R)"            "a'2(R)"           
#> [127] "AL-DC2(L)"         "SAD(-AMMC)"        "AL-DA4m(R)"       
#> [130] "a'L(L)"            "AL-VA5(R)"         "CX"               
#> [133] "PRW"               "FBl7"              "PB(L2)"           
#> [136] "AOTU(R)"           "PLP(R)"            "SPS(L)"           
#> [139] "EB"                "AL-DA4l(R)"        "PB(L3)"           
#> [142] "gL(R)"             "AL-VM1(R)"         "PB(R5)"           
#> [145] "g3(R)"             "ICL(L)"            "FBl1"             
#> [148] "LOP(R)"            "AL-DM1(L)"         "b1(R)"            
#> [151] "RUB(L)"            "AL-DL3(R)"         "AL-VM4(R)"        
#> [154] "AL-VM2(R)"         "ATL(L)"            "SIP(L)"           
#> [157] "b2(R)"             "PB(L4)"            "AL-DL4(R)"        
#> [160] "AL-VL1(R)"         "NO2(R)"            "LAL(L)"           
#> [163] "ICL(R)"            "mALT(L)"           "EBr2r4"           
#> [166] "AL-DM4(L)"         "AL-DP1m(L)"        "LX(R)"            
#> [169] "AL-DA3(L)"         "FBl4"              "AL-VP4(R)"        
#> [172] "AL-VP3(R)"         "PB(L8)"            "AL-DM6(L)"        
#> [175] "PB(R4)"            "AL-DC1(R)"         "b'2(R)"           
#> [178] "GC"                "AL-DL2d(R)"        "AL-DM5(R)"        
#> [181] "AL-DL2v(R)"        "AL-DP1m(R)"        "EBr6"             
#> [184] "PB"                "FBl6"              "AB(R)"            
#> [187] "LAL(R)"            "EBr1"              "VES(L)"           
#> [190] "LH(R)"             "PB(R2)"            "SAD"              
#> [193] "NO3(L)"            "INP"               "AL-DA3(R)"        
#> [196] "AL-DP1l(R)"        "VMNP"              "ATL(R)"           
#> [199] "SMP(R)"            "AL-VM5d(R)"        "mALT(R)"          
#> [202] "VES(R)"            "FBl3"              "SPS(R)"           
#> [205] "AL-DM6(R)"         "AL-DA2(L)"         "bL(L)"            
#> [208] "AL-DM3(R)"         "VLNP(R)"           "a1(R)"            
#> [211] "WED(R)"            "FBl9"              "EPA(L)"           
#> [214] "AL-DC3(R)"         "PB(L5)"            "SCL(R)"           
#> [217] "GNG"               "AL-DL1(R)"         "AL-VC3l(R)"       
#> [220] "AOT(R)"            "hemibrain"         "CRE(R)"           
#> [223] "a'1(R)"            "AL-VA4(R)"         "GF(R)"            
#> [226] "AL-DC4(L)"         "FLA(R)"            "AL-VA6(L)"        
#> [229] "PB(L6)"            "SNP(L)"            "AL-VA7l(R)"       
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "EPA(R)"            "FBl2"              "OL(R)"            
#>   [4] "a2(R)"             "EBr3d"             "IB"               
#>   [7] "AMMC"              "SPS(R)"            "PB(R7)"           
#>  [10] "CRE(R)"            "b2(R)"             "SCL(R)"           
#>  [13] "CA(R)"             "INP"               "PB(L8)"           
#>  [16] "a'L(R)"            "SIP(R)"            "FBl3"             
#>  [19] "aL(L)"             "EPA(L)"            "NO2(L)"           
#>  [22] "FLA(R)"            "g5(R)"             "LO(R)"            
#>  [25] "FBl5"              "mALT(L)"           "GF(R)"            
#>  [28] "PRW"               "GC"                "WED(R)"           
#>  [31] "g1(R)"             "NO1(R)"            "SAD(-AMMC)"       
#>  [34] "LAL(-GA)(R)"       "PB(R6)"            "PB(R8)"           
#>  [37] "SNP(R)"            "EB"                "gL(L)"            
#>  [40] "PB(L7)"            "PB(L2)"            "SAD"              
#>  [43] "NO1(L)"            "VMNP"              "PB(R9)"           
#>  [46] "NO(L)"             "SMP(L)"            "VES(R)"           
#>  [49] "FBl1"              "VES(L)"            "PB"               
#>  [52] "PB(R1)"            "CAN(R)"            "b'2(R)"           
#>  [55] "EBr5"              "PB(L6)"            "SNP(L)"           
#>  [58] "AB(R)"             "SCL(L)"            "GA(R)"            
#>  [61] "MB(L)"             "b'L(R)"            "g2(R)"            
#>  [64] "a'2(R)"            "BU(R)"             "NO3(L)"           
#>  [67] "MB(+ACA)(R)"       "PB(R5)"            "PENP"             
#>  [70] "aL(R)"             "RUB(L)"            "PB(L1)"           
#>  [73] "FBl6"              "MB(R)"             "SMP(R)"           
#>  [76] "EBr6"              "dACA(R)"           "g4(R)"            
#>  [79] "b'1(R)"            "AOTU(R)"           "PVLP(R)"          
#>  [82] "LH(R)"             "PB(R2)"            "b1(R)"            
#>  [85] "a3(R)"             "a'L(L)"            "ROB(R)"           
#>  [88] "ICL(L)"            "NO2(R)"            "PB(R4)"           
#>  [91] "EBr1"              "SPS(L)"            "PLP(R)"           
#>  [94] "NO(R)"             "FBl4"              "NO"               
#>  [97] "POC"               "EBr3am"            "PB(L5)"           
#> [100] "LX(R)"             "a'1(R)"            "PB(R3)"           
#> [103] "PB(L4)"            "VLNP(R)"           "RUB(R)"           
#> [106] "vACA(R)"           "a1(R)"             "CA(L)"            
#> [109] "gL(R)"             "ATL(R)"            "AVLP(R)"          
#> [112] "GNG"               "FBl7"              "GOR(R)"           
#> [115] "AOT(R)"            "CX"                "PB(L3)"           
#> [118] "NO3(R)"            "BU(L)"             "EBr2r4"           
#> [121] "bL(L)"             "LX(L)"             "bL(R)"            
#> [124] "AL(R)"             "FBl9"              "SLP(R)"           
#> [127] "mALT(R)"           "ME(R)"             "PED(R)"           
#> [130] "PB(L9)"            "LAL(R)"            "CRE(-ROB,-RUB)(R)"
#> [133] "CRE(L)"            "FBl8"              "IPS(R)"           
#> [136] "LAL(L)"            "g3(R)"             "hemibrain"        
#> [139] "EBr3pw"            "AB(L)"             "ATL(L)"           
#> [142] "SIP(L)"            "lACA(R)"           "GOR(L)"           
#> [145] "b'L(L)"            "LOP(R)"            "a'3(R)"           
#> [148] "FB"                "AL(L)"             "ICL(R)"           
#> [151] "AME(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "NO1(R)"            "AL-DC2(R)"         "AL-VA7m(R)"       
#>   [4] "AL-VL2p(R)"        "AL-VM1(R)"         "LAL(L)"           
#>   [7] "b2(R)"             "AL-D(L)"           "AL-VC1(R)"        
#>  [10] "AL-VC5(R)"         "FBl3"              "AB(L)"            
#>  [13] "b'2(R)"            "GOR(L)"            "a'L(R)"           
#>  [16] "AL(L)"             "FBl2"              "CRE(L)"           
#>  [19] "AL-VC2(R)"         "SIP(R)"            "AL-V(R)"          
#>  [22] "SMP(L)"            "PB(L3)"            "ICL(L)"           
#>  [25] "OL(R)"             "AL-DL3(R)"         "a2(R)"            
#>  [28] "AL-DM5(R)"         "hemibrain"         "a1(R)"            
#>  [31] "PB"                "PB(R1)"            "AL-VA6(R)"        
#>  [34] "SMP(R)"            "CAN(R)"            "AL-DL5(R)"        
#>  [37] "AL-VA4(R)"         "NO1(L)"            "AL-DM3(L)"        
#>  [40] "EBr3pw"            "LH(R)"             "NO3(R)"           
#>  [43] "b1(R)"             "a'L(L)"            "AL-DC3(R)"        
#>  [46] "AL-DC2(L)"         "FBl6"              "PB(R3)"           
#>  [49] "aL(R)"             "g2(R)"             "EBr1"             
#>  [52] "LX(R)"             "g4(R)"             "ICL(R)"           
#>  [55] "AL-DM4(L)"         "AL-D(R)"           "SIP(L)"           
#>  [58] "MB(L)"             "EB"                "EPA(L)"           
#>  [61] "ATL(L)"            "g1(R)"             "SCL(R)"           
#>  [64] "LO(R)"             "AL-VP1d(R)"        "SPS(R)"           
#>  [67] "CA(R)"             "AL-DA4l(R)"        "AL-DM2(R)"        
#>  [70] "AMMC"              "PRW"               "SNP(L)"           
#>  [73] "LX(L)"             "a3(R)"             "LOP(R)"           
#>  [76] "AL-DA1(R)"         "AL-VA1d(R)"        "CX"               
#>  [79] "CRE(R)"            "BU(R)"             "ROB(R)"           
#>  [82] "PVLP(R)"           "FBl9"              "PB(R7)"           
#>  [85] "PB(L6)"            "MB(+ACA)(R)"       "AL-DA4m(L)"       
#>  [88] "aL(L)"             "b'L(L)"            "RUB(R)"           
#>  [91] "NO"                "EBr3am"            "AL(R)"            
#>  [94] "AL-DM2(L)"         "AL-VM7v(L)"        "AL-VA7l(R)"       
#>  [97] "CRE(-ROB,-RUB)(R)" "AL-VM7v(R)"        "GA(R)"            
#> [100] "FBl5"              "AL-DC1(R)"         "VES(L)"           
#> [103] "AL-VL1(R)"         "EBr6"              "EBr3d"            
#> [106] "AL-DA2(L)"         "FBl4"              "ATL(R)"           
#> [109] "LAL(-GA)(R)"       "NO(R)"             "AL-VM2(R)"        
#> [112] "g3(R)"             "EPA(R)"            "NO3(L)"           
#> [115] "lACA(R)"           "FBl7"              "AL-VA2(R)"        
#> [118] "VMNP"              "bL(L)"             "LAL(R)"           
#> [121] "AL-VA5(R)"         "AL-VP5(R)"         "PENP"             
#> [124] "PB(R8)"            "AVLP(R)"           "NO(L)"            
#> [127] "AL-DM1(R)"         "mALT(L)"           "FB"               
#> [130] "b'L(R)"            "AL-VP4(R)"         "IB"               
#> [133] "CA(L)"             "PED(R)"            "PB(L9)"           
#> [136] "AL-VC4(R)"         "AL-VM5v(R)"        "AL-DP1m(R)"       
#> [139] "AL-VL2a(R)"        "GF(R)"             "AL-DL4(L)"        
#> [142] "PB(R9)"            "AL-DL2d(R)"        "AL-DL4(R)"        
#> [145] "AB(R)"             "IPS(R)"            "INP"              
#> [148] "AL-DA3(R)"         "AL-VA1v(R)"        "AL-DC1(L)"        
#> [151] "g5(R)"             "AL-DP1l(R)"        "NO2(L)"           
#> [154] "a'2(R)"            "SAD(-AMMC)"        "FBl8"             
#> [157] "RUB(L)"            "POC"               "b'1(R)"           
#> [160] "GC"                "dACA(R)"           "PB(R2)"           
#> [163] "MB(R)"             "AOT(R)"            "AL-DL1(R)"        
#> [166] "AL-VA3(R)"         "AL-VM7d(R)"        "PB(R6)"           
#> [169] "SNP(R)"            "gL(L)"             "AL-DC4(L)"        
#> [172] "NO2(R)"            "ME(R)"             "AL-DP1m(L)"       
#> [175] "AL-DA4m(R)"        "PB(L1)"            "VES(R)"           
#> [178] "GNG"               "AL-VC3m(R)"        "CRE(-RUB)(L)"     
#> [181] "VLNP(R)"           "AL-VM5d(R)"        "AL-VP1l(R)"       
#> [184] "AOTU(R)"           "PB(L7)"            "FB-column3"       
#> [187] "PB(L4)"            "AL-DM6(R)"         "PB(R4)"           
#> [190] "vACA(R)"           "BU(L)"             "AL-DL5(L)"        
#> [193] "AL-DM5(L)"         "AL-VM7d(L)"        "AME(R)"           
#> [196] "EBr5"              "FLA(R)"            "SLP(R)"           
#> [199] "AL-DC4(R)"         "PB(L2)"            "AL-VC3l(R)"       
#> [202] "AL-DM1(L)"         "a'1(R)"            "AL-DA2(R)"        
#> [205] "mALT(R)"           "AL-DM6(L)"         "FBl1"             
#> [208] "AL-VP1m(R)"        "a'3(R)"            "PLP(R)"           
#> [211] "AL-VP3(R)"         "PB(R5)"            "SCL(L)"           
#> [214] "GOR(R)"            "AL-DM3(R)"         "WED(R)"           
#> [217] "AL-DA3(L)"         "AL-VA6(L)"         "gL(R)"            
#> [220] "AL-VP2(R)"         "SAD"               "PB(L8)"           
#> [223] "SPS(L)"            "AL-VM4(R)"         "PB(L5)"           
#> [226] "EBr2r4"            "bL(R)"             "AL-DL2v(R)"       
#> [229] "AL-DM4(R)"         "AL-VM3(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "ICL(R)"            "PB(L4)"            "AL-VA6(R)"        
#>   [4] "AL-DM1(R)"         "PB(R6)"            "CRE(R)"           
#>   [7] "FB"                "SCL(R)"            "EBr2r4"           
#>  [10] "g1(R)"             "EBr6"              "ME(R)"            
#>  [13] "AL(L)"             "AL-D(R)"           "AL-VA1d(R)"       
#>  [16] "AL-VA7m(R)"        "NO(R)"             "CRE(L)"           
#>  [19] "AL-VC3m(R)"        "AL-VM7d(R)"        "RUB(L)"           
#>  [22] "PB(L3)"            "GA(R)"             "AL-D(L)"          
#>  [25] "AL-VA4(R)"         "a2(R)"             "AL-DA2(R)"        
#>  [28] "AL-DC2(R)"         "AL-DM3(R)"         "AL-DP1m(R)"       
#>  [31] "ROB(R)"            "AB(R)"             "AL-VA7l(R)"       
#>  [34] "BU(R)"             "AL-VC1(R)"         "PB(R5)"           
#>  [37] "AL-DP1m(L)"        "gL(R)"             "AL-VP2(R)"        
#>  [40] "IB"                "AL-VP5(R)"         "EPA(L)"           
#>  [43] "AL-DL2d(R)"        "AL-VL2a(R)"        "lACA(R)"          
#>  [46] "EPA(R)"            "AL-DL4(L)"         "NO"               
#>  [49] "a'1(R)"            "AL-DA4l(R)"        "BU(L)"            
#>  [52] "FBl8"              "CA(R)"             "AVLP(R)"          
#>  [55] "SLP(R)"            "AL-DA2(L)"         "FBl1"             
#>  [58] "AL-VA2(R)"         "AL-VM3(R)"         "CA(L)"            
#>  [61] "AL-DL5(R)"         "AL-VP1d(R)"        "EBr1"             
#>  [64] "AL-DL5(L)"         "ATL(R)"            "AL-DA4m(R)"       
#>  [67] "a'3(R)"            "GOR(L)"            "AL-DC2(L)"        
#>  [70] "FBl5"              "VLNP(R)"           "EBr3am"           
#>  [73] "AL-DA3(L)"         "PB(R3)"            "AB(L)"            
#>  [76] "hemibrain"         "AL-VM5d(R)"        "AOTU(R)"          
#>  [79] "AL-DM6(L)"         "bL(L)"             "mALT(L)"          
#>  [82] "AL-DC3(R)"         "AL-VP3(R)"         "AL-VP1m(R)"       
#>  [85] "PB(L2)"            "CRE(-RUB)(L)"      "a3(R)"            
#>  [88] "LH(R)"             "a'L(L)"            "a1(R)"            
#>  [91] "LAL(-GA)(R)"       "PB(R4)"            "AL-DM4(R)"        
#>  [94] "b'2(R)"            "SIP(L)"            "AL-DC4(L)"        
#>  [97] "AL-VA6(L)"         "AL-DL1(R)"         "b'1(R)"           
#> [100] "SAD"               "b2(R)"             "NO1(R)"           
#> [103] "AL-DC1(L)"         "FBl9"              "AMMC"             
#> [106] "aL(R)"             "AME(R)"            "AL(R)"            
#> [109] "AL-VC3l(R)"        "AL-VP4(R)"         "LO(R)"            
#> [112] "AL-DP1l(R)"        "AL-VL1(R)"         "AOT(R)"           
#> [115] "gL(L)"             "AL-VM7v(L)"        "SAD(-AMMC)"       
#> [118] "AL-DA1(R)"         "GF(R)"             "PB(L6)"           
#> [121] "AL-VM1(R)"         "NO2(L)"            "VES(R)"           
#> [124] "NO2(R)"            "OL(R)"             "AL-DL3(R)"        
#> [127] "AL-DM6(R)"         "PB(L1)"            "MB(+ACA)(R)"      
#> [130] "GOR(R)"            "PB(R7)"            "WED(R)"           
#> [133] "AL-DL2v(R)"        "FLA(R)"            "AL-DA4m(L)"       
#> [136] "AL-DM2(L)"         "CX"                "AL-VM7d(L)"       
#> [139] "AL-V(R)"           "MB(L)"             "a'L(R)"           
#> [142] "PVLP(R)"           "PB(L5)"            "LAL(L)"           
#> [145] "VES(L)"            "aL(L)"             "LX(R)"            
#> [148] "PB(R1)"            "FBl2"              "FBl4"             
#> [151] "IPS(R)"            "PENP"              "AL-VC4(R)"        
#> [154] "AL-DA3(R)"         "PED(R)"            "FB-column3"       
#> [157] "AL-DM5(R)"         "AL-VM4(R)"         "AL-VA3(R)"        
#> [160] "NO1(L)"            "EBr5"              "SPS(L)"           
#> [163] "SNP(R)"            "b1(R)"             "AL-DM5(L)"        
#> [166] "FBl7"              "SPS(R)"            "PB(L7)"           
#> [169] "SNP(L)"            "INP"               "AL-DC1(R)"        
#> [172] "AL-DM3(L)"         "GC"                "AL-DM1(L)"        
#> [175] "AL-VA5(R)"         "AL-VM5v(R)"        "PB"               
#> [178] "AL-VA1v(R)"        "AL-VM2(R)"         "LOP(R)"           
#> [181] "PB(L8)"            "PB(R8)"            "g4(R)"            
#> [184] "PB(L9)"            "b'L(L)"            "ICL(L)"           
#> [187] "AL-VM7v(R)"        "RUB(R)"            "vACA(R)"          
#> [190] "mALT(R)"           "PLP(R)"            "MB(R)"            
#> [193] "POC"               "FBl3"              "PB(R2)"           
#> [196] "g2(R)"             "PB(R9)"            "AL-DC4(R)"        
#> [199] "VMNP"              "SCL(L)"            "EBr3d"            
#> [202] "g5(R)"             "SMP(L)"            "AL-VL2p(R)"       
#> [205] "NO3(R)"            "EB"                "NO(L)"            
#> [208] "AL-VC2(R)"         "ATL(L)"            "AL-DM2(R)"        
#> [211] "PRW"               "SIP(R)"            "CAN(R)"           
#> [214] "AL-VP1l(R)"        "LX(L)"             "a'2(R)"           
#> [217] "AL-DM4(L)"         "EBr3pw"            "FBl6"             
#> [220] "CRE(-ROB,-RUB)(R)" "b'L(R)"            "dACA(R)"          
#> [223] "AL-DL4(R)"         "SMP(R)"            "g3(R)"            
#> [226] "NO3(L)"            "AL-VC5(R)"         "GNG"              
#> [229] "bL(R)"             "LAL(R)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method    path                                        
#> [1,] "GET"     "/api/cached/dailytype"                     
#> [2,] "GET"     "/api/dbmeta/database"                      
#> [3,] "POST"    "/api/v:ver/skeletons/skeleton/:dataset/:id"
#> [4,] "CONNECT" "/api/*"                                    
#> [5,] "GET"     "/api/cached/roicompleteness"               
#> [6,] "GET"     "/api/dbmeta/version"                       
#>      name                                                                                    
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getDailyType-fm"      
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatabase-fm"        
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [4,] "github.com/labstack/echo/v4.glob..func1"                                               
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROICompleteness-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getVersion-fm"         
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }