Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "SMP(R)"            "b2(R)"             "g2(R)"            
#>   [4] "AL(L)"             "LX(L)"             "LAL(R)"           
#>   [7] "VLNP(R)"           "EBr3d"             "FBl1"             
#>  [10] "CA(L)"             "PED(R)"            "ROB(R)"           
#>  [13] "bL(L)"             "FBl4"              "FBl5"             
#>  [16] "MB(R)"             "SMP(L)"            "FLA(R)"           
#>  [19] "EB"                "a'L(L)"            "VES(L)"           
#>  [22] "b'L(R)"            "NO2(R)"            "RUB(L)"           
#>  [25] "GNG"               "IB"                "SAD"              
#>  [28] "NO1(L)"            "FB"                "SIP(R)"           
#>  [31] "PVLP(R)"           "PENP"              "LAL(L)"           
#>  [34] "AOTU(R)"           "PB(R7)"            "g4(R)"            
#>  [37] "a'1(R)"            "EPA(R)"            "g5(R)"            
#>  [40] "AMMC"              "PB(L1)"            "PLP(R)"           
#>  [43] "AME(R)"            "lACA(R)"           "VMNP"             
#>  [46] "EBr2r4"            "MB(+ACA)(R)"       "LH(R)"            
#>  [49] "AOT(R)"            "CRE(L)"            "IPS(R)"           
#>  [52] "NO(L)"             "OL(R)"             "b'1(R)"           
#>  [55] "PB"                "VES(R)"            "FBl6"             
#>  [58] "NO2(L)"            "PB(L9)"            "ICL(R)"           
#>  [61] "POC"               "a2(R)"             "PB(R8)"           
#>  [64] "gL(L)"             "SNP(L)"            "SLP(R)"           
#>  [67] "PB(L5)"            "CX"                "LAL(-GA)(R)"      
#>  [70] "WED(R)"            "EBr5"              "BU(L)"            
#>  [73] "dACA(R)"           "PB(R3)"            "SCL(L)"           
#>  [76] "CRE(-ROB,-RUB)(R)" "mALT(R)"           "CRE(R)"           
#>  [79] "PB(L8)"            "a'3(R)"            "ICL(L)"           
#>  [82] "FBl3"              "aL(L)"             "SIP(L)"           
#>  [85] "EBr6"              "a'L(R)"            "NO"               
#>  [88] "bL(R)"             "SCL(R)"            "gL(R)"            
#>  [91] "NO3(L)"            "ATL(R)"            "INP"              
#>  [94] "PB(L2)"            "EBr1"              "a'2(R)"           
#>  [97] "FBl9"              "CA(R)"             "ME(R)"            
#> [100] "SNP(R)"            "PB(L3)"            "PRW"              
#> [103] "LX(R)"             "a3(R)"             "RUB(R)"           
#> [106] "PB(L6)"            "PB(R1)"            "PB(R2)"           
#> [109] "vACA(R)"           "PB(R5)"            "AB(L)"            
#> [112] "SPS(R)"            "PB(R9)"            "NO(R)"            
#> [115] "EBr3pw"            "AB(R)"             "PB(R4)"           
#> [118] "mALT(L)"           "PB(R6)"            "NO3(R)"           
#> [121] "b1(R)"             "EPA(L)"            "b'L(L)"           
#> [124] "ATL(L)"            "b'2(R)"            "g3(R)"            
#> [127] "GA(R)"             "hemibrain"         "SAD(-AMMC)"       
#> [130] "FBl2"              "GC"                "PB(L7)"           
#> [133] "FBl7"              "CAN(R)"            "PB(L4)"           
#> [136] "EBr3am"            "GOR(R)"            "AL(R)"            
#> [139] "GOR(L)"            "NO1(R)"            "SPS(L)"           
#> [142] "BU(R)"             "LO(R)"             "AVLP(R)"          
#> [145] "LOP(R)"            "g1(R)"             "aL(R)"            
#> [148] "MB(L)"             "GF(R)"             "FBl8"             
#> [151] "a1(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "PENP"              "a'L(L)"            "b'L(R)"           
#>   [4] "AL-DA3(L)"         "AL-DL2d(R)"        "LAL(-GA)(R)"      
#>   [7] "ME(R)"             "bL(L)"             "AL-DC2(R)"        
#>  [10] "mALT(L)"           "PB(R6)"            "a2(R)"            
#>  [13] "a'1(R)"            "g5(R)"             "ICL(L)"           
#>  [16] "PB(R3)"            "AL-DC3(R)"         "PB(L3)"           
#>  [19] "AL-VM7v(L)"        "CA(R)"             "LX(L)"            
#>  [22] "AL-VP1d(R)"        "FLA(R)"            "AOT(R)"           
#>  [25] "PB(L2)"            "NO(R)"             "PB(L7)"           
#>  [28] "AL-VA7m(R)"        "CX"                "LAL(L)"           
#>  [31] "FBl1"              "FBl5"              "a'3(R)"           
#>  [34] "aL(R)"             "VMNP"              "AL-VA1d(R)"       
#>  [37] "a1(R)"             "PB(R7)"            "PB(R4)"           
#>  [40] "AL-DM6(R)"         "VLNP(R)"           "OL(R)"            
#>  [43] "AL-VA7l(R)"        "AL-VC3l(R)"        "SAD"              
#>  [46] "EBr3d"             "AL-DC2(L)"         "CAN(R)"           
#>  [49] "INP"               "AL-VP1l(R)"        "vACA(R)"          
#>  [52] "AL-DL4(L)"         "PB(L4)"            "NO3(R)"           
#>  [55] "dACA(R)"           "LX(R)"             "AL-DM2(R)"        
#>  [58] "EBr1"              "AL-VP4(R)"         "POC"              
#>  [61] "AL-D(L)"           "AL-VM7v(R)"        "AL-VP3(R)"        
#>  [64] "SCL(L)"            "PB(L6)"            "NO(L)"            
#>  [67] "AL-DL4(R)"         "AL-VM2(R)"         "AL-DC4(L)"        
#>  [70] "AL(L)"             "AL-DM3(L)"         "PVLP(R)"          
#>  [73] "GF(R)"             "CRE(R)"            "LH(R)"            
#>  [76] "AL-DM1(R)"         "AL-VP2(R)"         "AL-VM4(R)"        
#>  [79] "SMP(R)"            "PED(R)"            "g4(R)"            
#>  [82] "PB(L9)"            "MB(+ACA)(R)"       "AL-DM4(R)"        
#>  [85] "ATL(R)"            "AL-DL1(R)"         "b'1(R)"           
#>  [88] "NO2(L)"            "PB(R8)"            "ROB(R)"           
#>  [91] "FBl4"              "FBl9"              "ATL(L)"           
#>  [94] "g1(R)"             "AL-DM2(L)"         "IPS(R)"           
#>  [97] "AL-DL2v(R)"        "PB(L1)"            "RUB(L)"           
#> [100] "MB(R)"             "AL-DA4m(R)"        "AL-VC3m(R)"       
#> [103] "FBl2"              "NO2(R)"            "a3(R)"            
#> [106] "SNP(L)"            "SIP(L)"            "AL-DA2(L)"        
#> [109] "AL-DM5(L)"         "AL-DM6(L)"         "AL-VM7d(L)"       
#> [112] "SMP(L)"            "PB"                "AL-VC4(R)"        
#> [115] "EBr3am"            "b1(R)"             "AL-VA2(R)"        
#> [118] "PB(R5)"            "SPS(R)"            "EPA(R)"           
#> [121] "AL-DC1(L)"         "AL-DC4(R)"         "LOP(R)"           
#> [124] "CRE(-RUB)(L)"      "AL-DM5(R)"         "AL-DP1l(R)"       
#> [127] "MB(L)"             "a'L(R)"            "AL-VM5v(R)"       
#> [130] "AMMC"              "WED(R)"            "PLP(R)"           
#> [133] "LAL(R)"            "AL-VL1(R)"         "b'2(R)"           
#> [136] "SNP(R)"            "AL-VA1v(R)"        "AL-VA3(R)"        
#> [139] "FB"                "PB(L8)"            "FBl8"             
#> [142] "RUB(R)"            "NO1(L)"            "AL-VM5d(R)"       
#> [145] "EB"                "AL-DL3(R)"         "gL(L)"            
#> [148] "FBl7"              "AL-VA4(R)"         "AL-VL2p(R)"       
#> [151] "g2(R)"             "SIP(R)"            "AL-DA1(R)"        
#> [154] "AL-DA2(R)"         "AL-DA4m(L)"        "gL(R)"            
#> [157] "SAD(-AMMC)"        "AL-VP1m(R)"        "CA(L)"            
#> [160] "GA(R)"             "FB-column3"        "SLP(R)"           
#> [163] "AL-VC1(R)"         "AL-VC5(R)"         "hemibrain"        
#> [166] "BU(L)"             "BU(R)"             "PB(R9)"           
#> [169] "AL-VL2a(R)"        "EBr5"              "CRE(L)"           
#> [172] "FBl3"              "g3(R)"             "AL-VA6(R)"        
#> [175] "AL-DM4(L)"         "AL-D(R)"           "NO"               
#> [178] "AME(R)"            "PB(R1)"            "GOR(R)"           
#> [181] "b'L(L)"            "AL-DM3(R)"         "SCL(R)"           
#> [184] "AL-VM1(R)"         "AL-VM3(R)"         "PRW"              
#> [187] "VES(R)"            "AL-V(R)"           "GOR(L)"           
#> [190] "mALT(R)"           "EBr6"              "PB(R2)"           
#> [193] "ICL(R)"            "AL-VM7d(R)"        "IB"               
#> [196] "AOTU(R)"           "EBr2r4"            "NO3(L)"           
#> [199] "bL(R)"             "AL-DP1m(R)"        "a'2(R)"           
#> [202] "AL-DL5(L)"         "AL-DC1(R)"         "FBl6"             
#> [205] "VES(L)"            "AB(L)"             "PB(L5)"           
#> [208] "CRE(-ROB,-RUB)(R)" "b2(R)"             "AL-DA3(R)"        
#> [211] "AL-DA4l(R)"        "AL-DL5(R)"         "GC"               
#> [214] "LO(R)"             "AB(R)"             "aL(L)"            
#> [217] "GNG"               "AL-VC2(R)"         "AL-DM1(L)"        
#> [220] "AL(R)"             "AL-VP5(R)"         "lACA(R)"          
#> [223] "AL-VA6(L)"         "EBr3pw"            "NO1(R)"           
#> [226] "AVLP(R)"           "SPS(L)"            "AL-DP1m(L)"       
#> [229] "EPA(L)"            "AL-VA5(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.2`
#> $`hemibrain:v1.2`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2`$ROIs
#>   [1] "MB(L)"             "FB-column3"        "FBl5"             
#>   [4] "AL(L)"             "NO"                "CRE(-RUB)(L)"     
#>   [7] "GA(R)"             "LX(R)"             "EBr2r4"           
#>  [10] "BU(L)"             "PB"                "SPS(R)"           
#>  [13] "LAL(-GA)(R)"       "AL-VC3l(R)"        "PENP"             
#>  [16] "CRE(-ROB,-RUB)(R)" "POC"               "WED(R)"           
#>  [19] "AL-DA3(L)"         "AL-VM7v(R)"        "g3(R)"            
#>  [22] "PB(R7)"            "FBl8"              "AL(R)"            
#>  [25] "AL-DL1(R)"         "g4(R)"             "EBr3pw"           
#>  [28] "EPA(L)"            "AL-DL4(R)"         "AL-VM3(R)"        
#>  [31] "PB(L8)"            "AL-DA2(L)"         "CRE(L)"           
#>  [34] "INP"               "mALT(R)"           "AL-DA4m(L)"       
#>  [37] "AL-DM4(L)"         "AL-VC2(R)"         "ATL(L)"           
#>  [40] "EPA(R)"            "AL-VA1v(R)"        "AL-VA3(R)"        
#>  [43] "AL-VC4(R)"         "PB(L2)"            "IB"               
#>  [46] "NO1(L)"            "AME(R)"            "GOR(R)"           
#>  [49] "AB(L)"             "SIP(L)"            "FBl7"             
#>  [52] "AL-VA2(R)"         "AL-DA4l(R)"        "SPS(L)"           
#>  [55] "a'L(L)"            "AL-DA1(R)"         "PB(R3)"           
#>  [58] "GNG"               "AL-VL1(R)"         "AL-VP4(R)"        
#>  [61] "g1(R)"             "gL(L)"             "ATL(R)"           
#>  [64] "AL-DM2(R)"         "AL-DM3(R)"         "PB(R5)"           
#>  [67] "ICL(R)"            "NO2(R)"            "AL-DC1(R)"        
#>  [70] "a'1(R)"            "AL-DM6(R)"         "GF(R)"            
#>  [73] "AL-DP1m(L)"        "PB(R9)"            "AOT(R)"           
#>  [76] "SMP(L)"            "ME(R)"             "AL-VM7v(L)"       
#>  [79] "AL-DM6(L)"         "AL-VL2p(R)"        "AL-VM7d(R)"       
#>  [82] "AL-DA3(R)"         "AL-VM4(R)"         "AOTU(R)"          
#>  [85] "NO(R)"             "PB(R4)"            "AL-VA6(L)"        
#>  [88] "AL-VC1(R)"         "AL-V(R)"           "SCL(L)"           
#>  [91] "SAD(-AMMC)"        "SLP(R)"            "AL-DL3(R)"        
#>  [94] "AL-DL5(R)"         "PB(L3)"            "FB"               
#>  [97] "CA(R)"             "EBr1"              "AL-DM3(L)"        
#> [100] "AL-VP1d(R)"        "g2(R)"             "ROB(R)"           
#> [103] "FBl1"              "AVLP(R)"           "a'L(R)"           
#> [106] "AL-DC4(L)"         "AL-D(L)"           "CA(L)"            
#> [109] "SNP(R)"            "PB(R1)"            "AL-VP5(R)"        
#> [112] "hemibrain"         "RUB(R)"            "LOP(R)"           
#> [115] "g5(R)"             "LH(R)"             "aL(L)"            
#> [118] "AL-VC3m(R)"        "NO2(L)"            "EB"               
#> [121] "MB(+ACA)(R)"       "PB(L4)"            "AL-VL2a(R)"       
#> [124] "PB(L1)"            "bL(L)"             "CAN(R)"           
#> [127] "NO(L)"             "PB(L5)"            "GOR(L)"           
#> [130] "bL(R)"             "a3(R)"             "PB(L9)"           
#> [133] "FBl2"              "AL-VA1d(R)"        "AL-VA4(R)"        
#> [136] "AL-VA7m(R)"        "a'3(R)"            "SAD"              
#> [139] "AL-DM5(L)"         "b'L(L)"            "LAL(L)"           
#> [142] "AL-VP2(R)"         "AL-DC2(L)"         "AL-VM7d(L)"       
#> [145] "RUB(L)"            "NO3(L)"            "FBl3"             
#> [148] "AL-DC4(R)"         "AL-DM4(R)"         "a2(R)"            
#> [151] "aL(R)"             "dACA(R)"           "AL-D(R)"          
#> [154] "GC"                "ICL(L)"            "AL-DA4m(R)"       
#> [157] "AL-DM1(R)"         "AL-VM5v(R)"        "AL-VP1m(R)"       
#> [160] "b'L(R)"            "EBr5"              "PVLP(R)"          
#> [163] "PLP(R)"            "AL-DM2(L)"         "AL-DC3(R)"        
#> [166] "BU(R)"             "AL-DP1m(R)"        "AL-VC5(R)"        
#> [169] "OL(R)"             "AL-VA5(R)"         "AL-VP1l(R)"       
#> [172] "PED(R)"            "VMNP"              "AL-DL5(L)"        
#> [175] "LO(R)"             "FBl4"              "LAL(R)"           
#> [178] "AL-DM5(R)"         "SCL(R)"            "AL-DL4(L)"        
#> [181] "gL(R)"             "IPS(R)"            "MB(R)"            
#> [184] "AL-DL2d(R)"        "AL-VA6(R)"         "AL-VM5d(R)"       
#> [187] "AL-VP3(R)"         "NO3(R)"            "VES(L)"           
#> [190] "AL-DC2(R)"         "AL-DL2v(R)"        "EBr3am"           
#> [193] "vACA(R)"           "PB(R6)"            "a'2(R)"           
#> [196] "PB(L7)"            "AMMC"              "FLA(R)"           
#> [199] "AL-DM1(L)"         "AL-VA7l(R)"        "PB(R2)"           
#> [202] "AL-VM2(R)"         "CX"                "PB(L6)"           
#> [205] "FBl6"              "AL-VM1(R)"         "CRE(R)"           
#> [208] "b2(R)"             "PB(R8)"            "b'2(R)"           
#> [211] "PRW"               "AL-DC1(L)"         "SNP(L)"           
#> [214] "a1(R)"             "b1(R)"             "AL-DP1l(R)"       
#> [217] "VLNP(R)"           "LX(L)"             "SIP(R)"           
#> [220] "FBl9"              "AB(R)"             "AL-DA2(R)"        
#> [223] "mALT(L)"           "SMP(R)"            "b'1(R)"           
#> [226] "lACA(R)"           "EBr6"              "NO1(R)"           
#> [229] "EBr3d"             "VES(R)"           
#> 
#> $`hemibrain:v1.2`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                
#> [1,] "GET"  "/api/dbmeta/version"               
#> [2,] "POST" "/api/npexplorer/simpleconnections" 
#> [3,] "POST" "/api/npexplorer/distribution"      
#> [4,] "GET"  "/api/cached/roiconnectivity"       
#> [5,] "GET"  "/api/npexplorer/roisinneuron"      
#> [6,] "POST" "/api/v:ver/npexplorer/distribution"
#>      name                                                                                             
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getVersion-fm"                  
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"     
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROIConnectivity-fm"         
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"     
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }