Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "SPS(R)"            "EBr5"              "EB"               
#>   [4] "PB(R1)"            "bL(L)"             "AL-VA6(R)"        
#>   [7] "AL-VP3(R)"         "a'1(R)"            "PB"               
#>  [10] "VLNP(R)"           "CRE(R)"            "SIP(R)"           
#>  [13] "aL(L)"             "a3(R)"             "AL-DM4(R)"        
#>  [16] "a2(R)"             "ME(R)"             "AL-DM5(L)"        
#>  [19] "AL-VM7v(L)"        "CA(R)"             "AL-DA4l(R)"       
#>  [22] "AL-DC1(R)"         "a'L(R)"            "AL-DM6(L)"        
#>  [25] "AL-DA2(R)"         "AL-DL2d(R)"        "ATL(L)"           
#>  [28] "PB(L7)"            "AL-DM2(R)"         "SMP(R)"           
#>  [31] "SCL(L)"            "a1(R)"             "LAL(L)"           
#>  [34] "PB(R7)"            "SCL(R)"            "SPS(L)"           
#>  [37] "ICL(L)"            "AL-VM5d(R)"        "VMNP"             
#>  [40] "SNP(R)"            "b'1(R)"            "MB(L)"            
#>  [43] "AL-DA1(R)"         "PB(L9)"            "LOP(R)"           
#>  [46] "SIP(L)"            "AL-DP1m(L)"        "WED(R)"           
#>  [49] "PB(R9)"            "FB-column3"        "AL-DP1l(R)"       
#>  [52] "GOR(R)"            "AL-VC3m(R)"        "ROB(R)"           
#>  [55] "gL(R)"             "PB(R3)"            "AL-VA1v(R)"       
#>  [58] "AL-VC1(R)"         "AL-VL2p(R)"        "aL(R)"            
#>  [61] "FB"                "AL-DM3(L)"         "FBl7"             
#>  [64] "AB(L)"             "BU(L)"             "FLA(R)"           
#>  [67] "AL-D(L)"           "ATL(R)"            "AL-DC3(R)"        
#>  [70] "g3(R)"             "RUB(L)"            "AL-VA4(R)"        
#>  [73] "RUB(R)"            "PB(R6)"            "NO1(L)"           
#>  [76] "NO1(R)"            "AL-DL5(L)"         "EBr3pw"           
#>  [79] "GC"                "b'L(R)"            "AL-DL4(R)"        
#>  [82] "AL-D(R)"           "AL-VA1d(R)"        "AL-VM3(R)"        
#>  [85] "AMMC"              "lACA(R)"           "LAL(-GA)(R)"      
#>  [88] "PLP(R)"            "AL-DA3(L)"         "a'L(L)"           
#>  [91] "g2(R)"             "PB(L8)"            "AL-DM6(R)"        
#>  [94] "AL-VC4(R)"         "AL-VM1(R)"         "mALT(R)"          
#>  [97] "AL-DM2(L)"         "FBl5"              "AL-VP1d(R)"       
#> [100] "EBr3am"            "NO2(L)"            "GA(R)"            
#> [103] "a'2(R)"            "PVLP(R)"           "AL-DP1m(R)"       
#> [106] "AL-VA5(R)"         "AL-DL4(L)"         "FBl2"             
#> [109] "FBl4"              "AL-VP2(R)"         "NO(R)"            
#> [112] "INP"               "NO3(R)"            "g4(R)"            
#> [115] "PB(R4)"            "AL-VC3l(R)"        "AL-VM7d(R)"       
#> [118] "AL-VM7v(R)"        "PB(L3)"            "PED(R)"           
#> [121] "LH(R)"             "EPA(L)"            "OL(R)"            
#> [124] "GF(R)"             "g1(R)"             "AME(R)"           
#> [127] "b2(R)"             "LX(L)"             "AL(R)"            
#> [130] "AL-VP5(R)"         "SAD"               "SNP(L)"           
#> [133] "LAL(R)"            "AL-DL5(R)"         "AL-DM1(R)"        
#> [136] "dACA(R)"           "NO"                "CRE(L)"           
#> [139] "bL(R)"             "SLP(R)"            "AL-VM2(R)"        
#> [142] "mALT(L)"           "hemibrain"         "EBr3d"            
#> [145] "AL-DA2(L)"         "SAD(-AMMC)"        "FBl9"             
#> [148] "AL-VA7l(R)"        "SMP(L)"            "EPA(R)"           
#> [151] "AL-VM5v(R)"        "AL-V(R)"           "PB(L1)"           
#> [154] "gL(L)"             "GNG"               "AL-DA3(R)"        
#> [157] "vACA(R)"           "IB112"             "FBl1"             
#> [160] "IPS(R)"            "VES(L)"            "AL-VC5(R)"        
#> [163] "AL(L)"             "a'3(R)"            "NO(L)"            
#> [166] "AL-DL1(R)"         "AL-VA7m(R)"        "AL-VL1(R)"        
#> [169] "AL-DC2(R)"         "IB"                "PB(L2)"           
#> [172] "PENP"              "PB(R8)"            "AVLP(R)"          
#> [175] "b'L(L)"            "CRE(-ROB,-RUB)(R)" "AL-DA4m(R)"       
#> [178] "PB(L4)"            "GOR(L)"            "AL-VM7d(L)"       
#> [181] "b1(R)"             "PB(L5)"            "VES(R)"           
#> [184] "AB(R)"             "LX(R)"             "CAN(R)"           
#> [187] "AL-DM5(R)"         "AL-VP4(R)"         "AOT(R)"           
#> [190] "FBl3"              "FBl8"              "AL-VA3(R)"        
#> [193] "PB(L6)"            "g5(R)"             "AL-DA4m(L)"       
#> [196] "AL-DC4(R)"         "AL-DC1(L)"         "AL-VA6(L)"        
#> [199] "NO2(R)"            "AL-DL2v(R)"        "AL-VC2(R)"        
#> [202] "CA(L)"             "AL-DM1(L)"         "ICL(R)"           
#> [205] "AL-VL2a(R)"        "EBr6"              "PRW"              
#> [208] "MB(+ACA)(R)"       "AL-VP1m(R)"        "PB(R5)"           
#> [211] "EBr2r4"            "EBr1"              "POC"              
#> [214] "b'2(R)"            "NO3(L)"            "AL-VP1l(R)"       
#> [217] "AL-DL3(R)"         "LO(R)"             "AL-DC2(L)"        
#> [220] "AL-DM4(L)"         "MB(R)"             "CX"               
#> [223] "AOTU(R)"           "BU(R)"             "CRE(-RUB)(L)"     
#> [226] "PB(R2)"            "AL-DM3(R)"         "AL-VA2(R)"        
#> [229] "AL-DC4(L)"         "FBl6"              "AL-VM4(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "g3(R)"             "VMNP"              "PB(R1)"           
#>   [4] "FBl3"              "PB(R2)"            "LOP(R)"           
#>   [7] "a'3(R)"            "PB(L3)"            "SAD"              
#>  [10] "aL(R)"             "EBr1"              "PB(L6)"           
#>  [13] "GNG"               "AB(L)"             "b'2(R)"           
#>  [16] "SNP(R)"            "PB(L9)"            "VES(R)"           
#>  [19] "LAL(R)"            "EBr3am"            "SMP(L)"           
#>  [22] "EBr3pw"            "ICL(L)"            "EBr5"             
#>  [25] "SAD(-AMMC)"        "b'L(L)"            "AL(R)"            
#>  [28] "dACA(R)"           "a'2(R)"            "LAL(L)"           
#>  [31] "g2(R)"             "g5(R)"             "EPA(L)"           
#>  [34] "OL(R)"             "VLNP(R)"           "AMMC"             
#>  [37] "PENP"              "EBr2r4"            "a'L(R)"           
#>  [40] "ROB(R)"            "CRE(L)"            "AME(R)"           
#>  [43] "AOT(R)"            "a3(R)"             "PB(L2)"           
#>  [46] "PED(R)"            "ME(R)"             "PB(R4)"           
#>  [49] "b'1(R)"            "AOTU(R)"           "PB(R9)"           
#>  [52] "gL(R)"             "FBl7"              "NO(L)"            
#>  [55] "GA(R)"             "g1(R)"             "EB"               
#>  [58] "NO(R)"             "FBl4"              "PB(R3)"           
#>  [61] "lACA(R)"           "AL(L)"             "FBl9"             
#>  [64] "ATL(L)"            "SIP(L)"            "GC"               
#>  [67] "BU(L)"             "EBr3d"             "EPA(R)"           
#>  [70] "NO"                "NO3(L)"            "MB(L)"            
#>  [73] "FBl5"              "PB(R6)"            "b2(R)"            
#>  [76] "PLP(R)"            "LH(R)"             "IPS(R)"           
#>  [79] "bL(R)"             "SMP(R)"            "GF(R)"            
#>  [82] "PB"                "aL(L)"             "NO1(L)"           
#>  [85] "AVLP(R)"           "mALT(R)"           "LX(R)"            
#>  [88] "a'L(L)"            "CX"                "CA(L)"            
#>  [91] "PRW"               "SCL(R)"            "RUB(L)"           
#>  [94] "LAL(-GA)(R)"       "b'L(R)"            "gL(L)"            
#>  [97] "LO(R)"             "FBl1"              "FBl6"             
#> [100] "PB(L4)"            "FB"                "SPS(L)"           
#> [103] "GOR(L)"            "WED(R)"            "MB(+ACA)(R)"      
#> [106] "MB(R)"             "vACA(R)"           "hemibrain"        
#> [109] "SPS(R)"            "PVLP(R)"           "FBl8"             
#> [112] "ATL(R)"            "IB"                "CRE(-ROB,-RUB)(R)"
#> [115] "INP"               "PB(R7)"            "EBr6"             
#> [118] "g4(R)"             "PB(L7)"            "CA(R)"            
#> [121] "b1(R)"             "CAN(R)"            "VES(L)"           
#> [124] "BU(R)"             "SLP(R)"            "PB(L5)"           
#> [127] "a2(R)"             "NO3(R)"            "PB(R8)"           
#> [130] "bL(L)"             "RUB(R)"            "CRE(R)"           
#> [133] "PB(L1)"            "NO1(R)"            "NO2(L)"           
#> [136] "FLA(R)"            "a'1(R)"            "SNP(L)"           
#> [139] "GOR(R)"            "PB(R5)"            "SCL(L)"           
#> [142] "a1(R)"             "SIP(R)"            "FBl2"             
#> [145] "AB(R)"             "POC"               "PB(L8)"           
#> [148] "ICL(R)"            "LX(L)"             "mALT(L)"          
#> [151] "NO2(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "NO1(R)"            "PB(L9)"            "AL-DP1m(R)"       
#>   [4] "b'L(R)"            "AL-VC3m(R)"        "EBr3am"           
#>   [7] "gL(R)"             "NO3(L)"            "AL-DL5(R)"        
#>  [10] "SMP(R)"            "GA(R)"             "AL(L)"            
#>  [13] "AL-VM7v(L)"        "PB(R2)"            "AVLP(R)"          
#>  [16] "FBl6"              "FBl8"              "GNG"              
#>  [19] "AL-VM1(R)"         "a2(R)"             "lACA(R)"          
#>  [22] "NO1(L)"            "EBr6"              "OL(R)"            
#>  [25] "AL-DM1(R)"         "LOP(R)"            "PB(R3)"           
#>  [28] "SMP(L)"            "AL-DM3(R)"         "AL-VL2a(R)"       
#>  [31] "ROB(R)"            "FBl2"              "NO2(R)"           
#>  [34] "AL-VC2(R)"         "SCL(L)"            "MB(+ACA)(R)"      
#>  [37] "vACA(R)"           "a'L(R)"            "a'L(L)"           
#>  [40] "MB(R)"             "POC"               "b'1(R)"           
#>  [43] "ATL(L)"            "AOTU(R)"           "SAD"              
#>  [46] "PRW"               "NO3(R)"            "FBl7"             
#>  [49] "AL-DL2d(R)"        "AL-VA3(R)"         "SCL(R)"           
#>  [52] "g5(R)"             "PVLP(R)"           "b1(R)"            
#>  [55] "LAL(R)"            "hemibrain"         "ME(R)"            
#>  [58] "SLP(R)"            "AL-VP3(R)"         "AL-VA1v(R)"       
#>  [61] "AL-VC1(R)"         "AL-VC3l(R)"        "AL-VM5v(R)"       
#>  [64] "EBr2r4"            "IPS(R)"            "AME(R)"           
#>  [67] "AL-DC4(R)"         "NO(R)"             "AL-VA4(R)"        
#>  [70] "NO"                "a'2(R)"            "CRE(-RUB)(L)"     
#>  [73] "AL-VC4(R)"         "AL-VM7v(R)"        "IB"               
#>  [76] "VMNP"              "EB"                "CRE(L)"           
#>  [79] "FBl4"              "NO(L)"             "AL-DL1(R)"        
#>  [82] "a'3(R)"            "FB-column3"        "AL-DM2(R)"        
#>  [85] "CRE(R)"            "BU(R)"             "FBl9"             
#>  [88] "AL(R)"             "PB(L2)"            "a1(R)"            
#>  [91] "CA(L)"             "PB(R9)"            "CA(R)"            
#>  [94] "AL-VP1l(R)"        "PED(R)"            "dACA(R)"          
#>  [97] "AL-DC4(L)"         "AL-DA3(R)"         "PB(L1)"           
#> [100] "AL-DM3(L)"         "AL-VM2(R)"         "AL-DA4m(L)"       
#> [103] "AL-DM4(L)"         "a3(R)"             "mALT(L)"          
#> [106] "EBr5"              "PB(L7)"            "PB(R6)"           
#> [109] "FB"                "EPA(R)"            "PB(L5)"           
#> [112] "NO2(L)"            "AL-DM6(L)"         "b2(R)"            
#> [115] "FLA(R)"            "AL-DM5(L)"         "PB(R4)"           
#> [118] "PB(L4)"            "VES(L)"            "LX(R)"            
#> [121] "FBl1"              "PB(R5)"            "mALT(R)"          
#> [124] "PB(L8)"            "SNP(R)"            "AL-DA1(R)"        
#> [127] "AL-VP5(R)"         "aL(R)"             "EBr3pw"           
#> [130] "VLNP(R)"           "g3(R)"             "g1(R)"            
#> [133] "b'L(L)"            "AL-V(R)"           "AL-VP4(R)"        
#> [136] "AL-VP1d(R)"        "AL-VL1(R)"         "AL-VM5d(R)"       
#> [139] "PB"                "PB(L6)"            "AL-VL2p(R)"       
#> [142] "CRE(-ROB,-RUB)(R)" "AL-DC2(L)"         "AL-DL4(L)"        
#> [145] "AL-VA6(L)"         "AB(R)"             "AL-DP1m(L)"       
#> [148] "SPS(R)"            "LAL(-GA)(R)"       "AL-DM1(L)"        
#> [151] "AL-VM3(R)"         "AL-DA2(L)"         "GOR(R)"           
#> [154] "AL-DL4(R)"         "LAL(L)"            "EPA(L)"           
#> [157] "AL-DC2(R)"         "CX"                "gL(L)"            
#> [160] "AL-DL5(L)"         "AL-VM7d(L)"        "AL-DM4(R)"        
#> [163] "EBr3d"             "AL-D(L)"           "AL-VA7m(R)"       
#> [166] "PLP(R)"            "aL(L)"             "AL-DA4l(R)"       
#> [169] "PENP"              "MB(L)"             "SIP(R)"           
#> [172] "ICL(L)"            "RUB(L)"            "AL-DM6(R)"        
#> [175] "AL-VA6(R)"         "CAN(R)"            "GC"               
#> [178] "AL-DA3(L)"         "VES(R)"            "AL-DC1(R)"        
#> [181] "PB(R1)"            "LX(L)"             "bL(R)"            
#> [184] "PB(L3)"            "GOR(L)"            "AL-DC1(L)"        
#> [187] "INP"               "g2(R)"             "LH(R)"            
#> [190] "AL-DA4m(R)"        "AL-DL3(R)"         "AL-D(R)"          
#> [193] "AL-VP2(R)"         "SNP(L)"            "AL-VA5(R)"        
#> [196] "AB(L)"             "AL-DM2(L)"         "AL-DC3(R)"        
#> [199] "AL-VA7l(R)"        "EBr1"              "bL(L)"            
#> [202] "AL-VM7d(R)"        "WED(R)"            "AL-DL2v(R)"       
#> [205] "AL-VA1d(R)"        "b'2(R)"            "SIP(L)"           
#> [208] "a'1(R)"            "BU(L)"             "FBl5"             
#> [211] "AL-DM5(R)"         "AOT(R)"            "ICL(R)"           
#> [214] "PB(R7)"            "SAD(-AMMC)"        "LO(R)"            
#> [217] "GF(R)"             "AL-DP1l(R)"        "AL-VA2(R)"        
#> [220] "AL-VM4(R)"         "SPS(L)"            "FBl3"             
#> [223] "AL-VC5(R)"         "RUB(R)"            "AMMC"             
#> [226] "PB(R8)"            "g4(R)"             "ATL(R)"           
#> [229] "AL-DA2(R)"         "AL-VP1m(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "EBr6"              "g5(R)"             "FBl3"             
#>   [4] "PB(L2)"            "aL(L)"             "PB(R4)"           
#>   [7] "GOR(R)"            "AL-VM7v(R)"        "AL-VM7d(L)"       
#>  [10] "AL-DL1(R)"         "LX(L)"             "AL-DM4(R)"        
#>  [13] "AL-VM7d(R)"        "AL-DA4m(L)"        "PB(R2)"           
#>  [16] "POC"               "IB"                "GF(R)"            
#>  [19] "FBl2"              "LH(R)"             "AOT(R)"           
#>  [22] "AL-DP1m(L)"        "NO(R)"             "IPS(R)"           
#>  [25] "AL-VP2(R)"         "AL-DM3(R)"         "AL-VL2a(R)"       
#>  [28] "RUB(R)"            "AL-VC3m(R)"        "AL-DM3(L)"        
#>  [31] "RUB(L)"            "AL(R)"             "AL-DL2d(R)"       
#>  [34] "LX(R)"             "BU(R)"             "FBl9"             
#>  [37] "ATL(L)"            "PB(L1)"            "PB(L8)"           
#>  [40] "PLP(R)"            "AL-VM4(R)"         "SIP(L)"           
#>  [43] "mALT(R)"           "FLA(R)"            "SAD(-AMMC)"       
#>  [46] "a'2(R)"            "VLNP(R)"           "PB(L3)"           
#>  [49] "PRW"               "PB(R7)"            "PB"               
#>  [52] "AL-VL1(R)"         "a1(R)"             "SPS(R)"           
#>  [55] "SCL(R)"            "VES(R)"            "SLP(R)"           
#>  [58] "AB(L)"             "AB(R)"             "bL(R)"            
#>  [61] "AL-DA4l(R)"        "AL-DP1l(R)"        "AL-DA3(L)"        
#>  [64] "ATL(R)"            "b'L(L)"            "AL-VA4(R)"        
#>  [67] "hemibrain"         "AL-DL4(L)"         "AL-D(R)"          
#>  [70] "AL-VA6(R)"         "LAL(L)"            "b2(R)"            
#>  [73] "EBr5"              "AL-VM3(R)"         "AL-VP3(R)"        
#>  [76] "FB"                "b'L(R)"            "AMMC"             
#>  [79] "GC"                "ROB(R)"            "gL(R)"            
#>  [82] "FBl6"              "AME(R)"            "AL-VC4(R)"        
#>  [85] "EPA(R)"            "AL-DL3(R)"         "AL-VC5(R)"        
#>  [88] "PENP"              "GA(R)"             "SPS(L)"           
#>  [91] "a'L(R)"            "AL-VA1d(R)"        "LOP(R)"           
#>  [94] "GOR(L)"            "PB(L7)"            "GNG"              
#>  [97] "CRE(R)"            "dACA(R)"           "NO1(R)"           
#> [100] "EBr1"              "AL-DC3(R)"         "AL-VA5(R)"        
#> [103] "lACA(R)"           "EBr3am"            "AL-DP1m(R)"       
#> [106] "ME(R)"             "MB(+ACA)(R)"       "AL-DL2v(R)"       
#> [109] "AL-DL4(R)"         "BU(L)"             "ICL(R)"           
#> [112] "NO2(R)"            "VES(L)"            "AL-DA1(R)"        
#> [115] "AL-VA1v(R)"        "AL-VA3(R)"         "FBl4"             
#> [118] "FBl7"              "AL-DM1(R)"         "CA(L)"            
#> [121] "g4(R)"             "LAL(R)"            "AL-VA7m(R)"       
#> [124] "AL-VM1(R)"         "SNP(L)"            "ICL(L)"           
#> [127] "NO(L)"             "b'1(R)"            "AL-D(L)"          
#> [130] "PB(L6)"            "AL-DC1(R)"         "AL-VP1l(R)"       
#> [133] "AL-VL2p(R)"        "PB(R6)"            "b'2(R)"           
#> [136] "AL-DC2(L)"         "AL-DA3(R)"         "EPA(L)"           
#> [139] "a3(R)"             "a'L(L)"            "AL-VM2(R)"        
#> [142] "WED(R)"            "NO3(R)"            "AL(L)"            
#> [145] "AVLP(R)"           "AL-VP4(R)"         "PVLP(R)"          
#> [148] "AL-VC3l(R)"        "mALT(L)"           "AL-DM6(L)"        
#> [151] "AL-VM7v(L)"        "AL-VC1(R)"         "SMP(L)"           
#> [154] "PB(R3)"            "CAN(R)"            "AL-DM5(R)"        
#> [157] "CA(R)"             "AL-VA7l(R)"        "AL-DM2(R)"        
#> [160] "AL-VP5(R)"         "g2(R)"             "AL-DM5(L)"        
#> [163] "PB(R1)"            "EBr3pw"            "g1(R)"            
#> [166] "FBl5"              "AL-VC2(R)"         "a2(R)"            
#> [169] "NO1(L)"            "AL-DM1(L)"         "CRE(L)"           
#> [172] "EBr2r4"            "AL-DC1(L)"         "AL-DL5(L)"        
#> [175] "MB(R)"             "a'1(R)"            "bL(L)"            
#> [178] "b1(R)"             "aL(R)"             "PB(R8)"           
#> [181] "AL-DC4(R)"         "SCL(L)"            "FBl8"             
#> [184] "a'3(R)"            "AL-DA2(L)"         "AL-DM2(L)"        
#> [187] "AL-DM4(L)"         "AL-DC4(L)"         "AL-DC2(R)"        
#> [190] "AL-VM5d(R)"        "MB(L)"             "EBr3d"            
#> [193] "CRE(-ROB,-RUB)(R)" "SAD"               "LO(R)"            
#> [196] "AL-V(R)"           "PB(L5)"            "LAL(-GA)(R)"      
#> [199] "SNP(R)"            "VMNP"              "NO"               
#> [202] "AL-VM5v(R)"        "NO3(L)"            "OL(R)"            
#> [205] "AL-VA2(R)"         "NO2(L)"            "INP"              
#> [208] "PB(L4)"            "PB(R5)"            "vACA(R)"          
#> [211] "EB"                "AL-VA6(L)"         "CRE(-RUB)(L)"     
#> [214] "FBl1"              "gL(L)"             "PB(R9)"           
#> [217] "FB-column3"        "CX"                "AL-DA4m(R)"       
#> [220] "AL-DM6(R)"         "AL-VP1m(R)"        "AOTU(R)"          
#> [223] "SIP(R)"            "PB(L9)"            "AL-DA2(R)"        
#> [226] "SMP(R)"            "g3(R)"             "PED(R)"           
#> [229] "AL-DL5(R)"         "AL-VP1d(R)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method  path                                     
#> [1,] "TRACE" "/api/*"                                 
#> [2,] "POST"  "/api/skeletons/skeleton/:dataset/:id"   
#> [3,] "GET"   "/api/v:ver/roimeshes/mesh/:dataset/:roi"
#> [4,] "GET"   "/api/cached/roiconnectivity"            
#> [5,] "GET"   "/api/v:ver/dbmeta/datasets"             
#> [6,] "POST"  "/api/*"                                 
#>      name                                                                                    
#> [1,] "github.com/labstack/echo.glob..func1"                                                  
#> [2,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROIConnectivity-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [6,] "github.com/labstack/echo.glob..func1"                                                  
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }