Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "PENP"              "CRE(R)"            "AL-VM7v(L)"       
#>   [4] "FBl7"              "AL-VC3m(R)"        "LAL(L)"           
#>   [7] "lACA(R)"           "AB(L)"             "IPS(R)"           
#>  [10] "AL-VP1d(R)"        "WED(R)"            "BU(L)"            
#>  [13] "MB(R)"             "AL-DA4m(R)"        "AL-VC5(R)"        
#>  [16] "AL-VA3(R)"         "dACA(R)"           "LAL(R)"           
#>  [19] "AL-DC1(R)"         "AL-DC4(R)"         "RUB(R)"           
#>  [22] "AL-VM7v(R)"        "a'3(R)"            "CRE(-ROB,-RUB)(R)"
#>  [25] "PB(R1)"            "FBl1"              "a3(R)"            
#>  [28] "AL-DA4l(R)"        "PB(R6)"            "hemibrain"        
#>  [31] "INP"               "SLP(R)"            "AL-VA7m(R)"       
#>  [34] "a'2(R)"            "FBl5"              "AL-VC1(R)"        
#>  [37] "AL-VM3(R)"         "SPS(R)"            "EBr1"             
#>  [40] "AL-DM2(L)"         "g5(R)"             "CA(R)"            
#>  [43] "AL-DL2v(R)"        "AL-DL5(R)"         "AL-VM7d(R)"       
#>  [46] "AL-DA2(L)"         "AL-DM5(L)"         "NO"               
#>  [49] "AL-VA2(R)"         "AL-VA4(R)"         "PB(L2)"           
#>  [52] "AL-DL4(L)"         "PB(R9)"            "FBl6"             
#>  [55] "mALT(L)"           "NO1(L)"            "SPS(L)"           
#>  [58] "RUB(L)"            "AL-VA7l(R)"        "AOTU(R)"          
#>  [61] "PLP(R)"            "AL-DC1(L)"         "AL-DP1m(L)"       
#>  [64] "AL-VL2a(R)"        "VLNP(R)"           "ATL(L)"           
#>  [67] "PRW"               "BU(R)"             "AL-DC4(L)"        
#>  [70] "AL-DC2(R)"         "AL-DM2(R)"         "NO3(R)"           
#>  [73] "PB(L7)"            "AL(R)"             "EBr5"             
#>  [76] "FBl8"              "a'1(R)"            "AL-VM2(R)"        
#>  [79] "AL-VM7d(L)"        "a'L(L)"            "OL(R)"            
#>  [82] "AL-DM3(R)"         "SNP(R)"            "AL-VC3l(R)"       
#>  [85] "b'1(R)"            "GOR(L)"            "PB(L4)"           
#>  [88] "AL-VP4(R)"         "AL-D(L)"           "NO2(R)"           
#>  [91] "PB"                "SIP(R)"            "AL-DA3(R)"        
#>  [94] "AL-DC3(R)"         "SMP(R)"            "AB(R)"            
#>  [97] "AL-VL1(R)"         "NO2(L)"            "g1(R)"            
#> [100] "PB(R2)"            "ICL(R)"            "b1(R)"            
#> [103] "AL-VP2(R)"         "SAD"               "g2(R)"            
#> [106] "PB(R4)"            "FBl3"              "EBr2r4"           
#> [109] "PB(L9)"            "SAD(-AMMC)"        "POC"              
#> [112] "SMP(L)"            "PB(R7)"            "bL(L)"            
#> [115] "VES(L)"            "b2(R)"             "EB"               
#> [118] "FB-column3"        "EPA(L)"            "NO(L)"            
#> [121] "AL-DL2d(R)"        "CX"                "aL(R)"            
#> [124] "AL(L)"             "LX(R)"             "EBr3pw"           
#> [127] "AL-DL4(R)"         "AL-VA5(R)"         "EBr3am"           
#> [130] "IB112"             "FBl9"              "AL-VC2(R)"        
#> [133] "gL(L)"             "ATL(R)"            "bL(R)"            
#> [136] "AL-VP1m(R)"        "AL-D(R)"           "PB(R5)"           
#> [139] "FLA(R)"            "NO(R)"             "PVLP(R)"          
#> [142] "PB(R3)"            "AL-DA1(R)"         "PED(R)"           
#> [145] "ROB(R)"            "PB(L6)"            "AL-VL2p(R)"       
#> [148] "AVLP(R)"           "b'L(L)"            "AL-VC4(R)"        
#> [151] "LOP(R)"            "AL-DM3(L)"         "AL-DM6(R)"        
#> [154] "AL-DP1l(R)"        "AL-VM5v(R)"        "b'2(R)"           
#> [157] "GC"                "b'L(R)"            "EBr6"             
#> [160] "FBl4"              "AL-DM4(R)"         "SCL(L)"           
#> [163] "a2(R)"             "MB(L)"             "CRE(L)"           
#> [166] "AL-DA2(R)"         "AL-VP5(R)"         "SIP(L)"           
#> [169] "NO1(R)"            "AL-DL3(R)"         "PB(L5)"           
#> [172] "PB(L3)"            "PB(R8)"            "AL-DC2(L)"        
#> [175] "AL-DL5(L)"         "AL-VA6(L)"         "aL(L)"            
#> [178] "AL-VA1d(R)"        "CA(L)"             "FB"               
#> [181] "LX(L)"             "IB"                "LAL(-GA)(R)"      
#> [184] "AL-DM1(L)"         "AL-VA6(R)"         "FBl2"             
#> [187] "vACA(R)"           "AL-DA4m(L)"        "EPA(R)"           
#> [190] "AL-DM6(L)"         "MB(+ACA)(R)"       "AL-DL1(R)"        
#> [193] "AOT(R)"            "SCL(R)"            "EBr3d"            
#> [196] "NO3(L)"            "AME(R)"            "AL-VA1v(R)"       
#> [199] "PB(L1)"            "PB(L8)"            "a'L(R)"           
#> [202] "g4(R)"             "LO(R)"             "AL-VM1(R)"        
#> [205] "GA(R)"             "VMNP"              "gL(R)"            
#> [208] "SNP(L)"            "g3(R)"             "ICL(L)"           
#> [211] "AL-DM5(R)"         "AL-VM4(R)"         "AL-VP3(R)"        
#> [214] "GF(R)"             "AL-DM4(L)"         "VES(R)"           
#> [217] "ME(R)"             "GNG"               "AL-DM1(R)"        
#> [220] "AL-V(R)"           "AL-VP1l(R)"        "AMMC"             
#> [223] "AL-VM5d(R)"        "mALT(R)"           "AL-DA3(L)"        
#> [226] "CRE(-RUB)(L)"      "LH(R)"             "CAN(R)"           
#> [229] "GOR(R)"            "AL-DP1m(R)"        "a1(R)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "a'1(R)"            "FBl8"              "WED(R)"           
#>   [4] "EBr3pw"            "CAN(R)"            "FBl1"             
#>   [7] "b1(R)"             "AL(R)"             "POC"              
#>  [10] "b'2(R)"            "GA(R)"             "b2(R)"            
#>  [13] "PLP(R)"            "PB(R9)"            "ATL(R)"           
#>  [16] "a'L(L)"            "SMP(L)"            "PB(L3)"           
#>  [19] "LAL(-GA)(R)"       "PB(R7)"            "g2(R)"            
#>  [22] "FBl5"              "IPS(R)"            "AME(R)"           
#>  [25] "SAD"               "b'L(R)"            "PB(L7)"           
#>  [28] "ICL(R)"            "LOP(R)"            "CX"               
#>  [31] "INP"               "GOR(R)"            "VES(L)"           
#>  [34] "IB"                "SCL(L)"            "NO2(L)"           
#>  [37] "MB(L)"             "PB(R8)"            "NO3(R)"           
#>  [40] "LX(L)"             "NO(L)"             "a'3(R)"           
#>  [43] "EBr2r4"            "NO"                "bL(L)"            
#>  [46] "a1(R)"             "g3(R)"             "SPS(R)"           
#>  [49] "EPA(R)"            "CRE(L)"            "FBl3"             
#>  [52] "GNG"               "lACA(R)"           "mALT(R)"          
#>  [55] "LO(R)"             "AB(R)"             "OL(R)"            
#>  [58] "PRW"               "aL(R)"             "EBr1"             
#>  [61] "LX(R)"             "EBr3am"            "hemibrain"        
#>  [64] "g4(R)"             "AL(L)"             "VES(R)"           
#>  [67] "PVLP(R)"           "FBl2"              "PENP"             
#>  [70] "AOTU(R)"           "g1(R)"             "BU(R)"            
#>  [73] "PB(R5)"            "SIP(L)"            "PB(R4)"           
#>  [76] "mALT(L)"           "ME(R)"             "EB"               
#>  [79] "FBl4"              "LH(R)"             "SMP(R)"           
#>  [82] "GF(R)"             "GC"                "BU(L)"            
#>  [85] "CA(L)"             "GOR(L)"            "NO1(L)"           
#>  [88] "PB(L6)"            "SNP(L)"            "LAL(R)"           
#>  [91] "PB(L4)"            "ATL(L)"            "SAD(-AMMC)"       
#>  [94] "PB(R2)"            "AVLP(R)"           "b'1(R)"           
#>  [97] "CRE(-ROB,-RUB)(R)" "VLNP(R)"           "FB"               
#> [100] "VMNP"              "SIP(R)"            "gL(R)"            
#> [103] "MB(+ACA)(R)"       "FBl6"              "PB(R3)"           
#> [106] "PB(L9)"            "CA(R)"             "SLP(R)"           
#> [109] "PB(L1)"            "NO(R)"             "RUB(R)"           
#> [112] "a3(R)"             "PED(R)"            "NO1(R)"           
#> [115] "SPS(L)"            "PB"                "ICL(L)"           
#> [118] "FBl7"              "RUB(L)"            "CRE(R)"           
#> [121] "dACA(R)"           "PB(L8)"            "a'L(R)"           
#> [124] "NO3(L)"            "bL(R)"             "vACA(R)"          
#> [127] "PB(R6)"            "EBr6"              "FBl9"             
#> [130] "NO2(R)"            "MB(R)"             "EBr3d"            
#> [133] "gL(L)"             "g5(R)"             "EPA(L)"           
#> [136] "a'2(R)"            "SNP(R)"            "ROB(R)"           
#> [139] "b'L(L)"            "AOT(R)"            "PB(R1)"           
#> [142] "aL(L)"             "PB(L5)"            "AMMC"             
#> [145] "LAL(L)"            "SCL(R)"            "AB(L)"            
#> [148] "a2(R)"             "EBr5"              "FLA(R)"           
#> [151] "PB(L2)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DC2(L)"         "RUB(L)"            "PB(R4)"           
#>   [4] "SLP(R)"            "PB(R5)"            "FB"               
#>   [7] "ME(R)"             "AL-DA1(R)"         "PRW"              
#>  [10] "WED(R)"            "NO1(L)"            "CRE(L)"           
#>  [13] "AL-V(R)"           "CA(L)"             "NO2(L)"           
#>  [16] "a'L(R)"            "AL-VM2(R)"         "LAL(-GA)(R)"      
#>  [19] "PB(L8)"            "EPA(R)"            "AL-DM1(R)"        
#>  [22] "SAD"               "CRE(R)"            "PB(R8)"           
#>  [25] "AL-DM1(L)"         "bL(R)"             "AL-DL3(R)"        
#>  [28] "LAL(L)"            "MB(L)"             "g2(R)"            
#>  [31] "SPS(L)"            "AL-DM3(L)"         "AL-DM4(R)"        
#>  [34] "AL-VP1d(R)"        "AOTU(R)"           "AL-DP1m(L)"       
#>  [37] "EPA(L)"            "VES(L)"            "AL-VM7d(R)"       
#>  [40] "PB(L5)"            "hemibrain"         "SIP(R)"           
#>  [43] "CAN(R)"            "AL-DL2v(R)"        "AL-DP1l(R)"       
#>  [46] "AL-VA7l(R)"        "AL-VP1l(R)"        "PED(R)"           
#>  [49] "mALT(R)"           "NO"                "gL(R)"            
#>  [52] "AL-DA4m(R)"        "VLNP(R)"           "PB(R7)"           
#>  [55] "LX(R)"             "BU(L)"             "FBl1"             
#>  [58] "AB(R)"             "NO2(R)"            "GA(R)"            
#>  [61] "GC"                "EBr3pw"            "GOR(R)"           
#>  [64] "lACA(R)"           "AL-DM5(L)"         "LAL(R)"           
#>  [67] "AL(R)"             "AL-DM3(R)"         "AL-VA3(R)"        
#>  [70] "AL-VA5(R)"         "dACA(R)"           "AL-DM6(L)"        
#>  [73] "AL-VM5d(R)"        "FLA(R)"            "a'2(R)"           
#>  [76] "AL-D(L)"           "PB(L9)"            "b1(R)"            
#>  [79] "AL-DC3(R)"         "AL-DL1(R)"         "AL-VM7v(R)"       
#>  [82] "AL-DM4(L)"         "VES(R)"            "a3(R)"            
#>  [85] "AL-DM2(R)"         "GOR(L)"            "ROB(R)"           
#>  [88] "ATL(R)"            "AL-DL2d(R)"        "AL(L)"            
#>  [91] "AL-DA2(L)"         "CA(R)"             "AL-DM5(R)"        
#>  [94] "AMMC"              "FB-column3"        "AL-DM6(R)"        
#>  [97] "AL-VC3m(R)"        "AL-VP4(R)"         "CRE(-ROB,-RUB)(R)"
#> [100] "PB(L3)"            "PB(R1)"            "aL(L)"            
#> [103] "NO(R)"             "AME(R)"            "MB(R)"            
#> [106] "AL-VP2(R)"         "EBr3d"             "FBl9"             
#> [109] "AL-VA2(R)"         "AL-VL2a(R)"        "GF(R)"            
#> [112] "AL-VM7d(L)"        "AL-VM7v(L)"        "FBl4"             
#> [115] "FBl5"              "CX"                "AL-DA4m(L)"       
#> [118] "FBl2"              "a'L(L)"            "AL-VL2p(R)"       
#> [121] "PB(R6)"            "FBl8"              "LH(R)"            
#> [124] "SNP(R)"            "SNP(L)"            "FBl7"             
#> [127] "OL(R)"             "AL-VM1(R)"         "AL-VM4(R)"        
#> [130] "gL(L)"             "AL-DM2(L)"         "PB"               
#> [133] "a'1(R)"            "PB(R2)"            "AL-VP3(R)"        
#> [136] "AB(L)"             "VMNP"              "AL-DC1(R)"        
#> [139] "aL(R)"             "PB(L1)"            "NO1(R)"           
#> [142] "LX(L)"             "AL-DA3(R)"         "AL-VM3(R)"        
#> [145] "g1(R)"             "EB"                "AL-DL5(L)"        
#> [148] "AL-VP1m(R)"        "PENP"              "a2(R)"            
#> [151] "b'2(R)"            "SIP(L)"            "AL-DL4(L)"        
#> [154] "INP"               "AL-VA6(R)"         "AL-DL5(R)"        
#> [157] "AL-VC1(R)"         "SCL(R)"            "ICL(L)"           
#> [160] "PB(L4)"            "b'1(R)"            "LOP(R)"           
#> [163] "bL(L)"             "MB(+ACA)(R)"       "AL-VA7m(R)"       
#> [166] "mALT(L)"           "EBr1"              "PB(R9)"           
#> [169] "AL-VA1d(R)"        "AL-VL1(R)"         "AL-DC4(R)"        
#> [172] "RUB(R)"            "SCL(L)"            "BU(R)"            
#> [175] "PVLP(R)"           "AL-DL4(R)"         "AL-VM5v(R)"       
#> [178] "ATL(L)"            "SPS(R)"            "EBr6"             
#> [181] "AL-DC4(L)"         "PB(L7)"            "AL-VC4(R)"        
#> [184] "AL-VC5(R)"         "AOT(R)"            "g5(R)"            
#> [187] "AL-D(R)"           "AL-VC3l(R)"        "a1(R)"            
#> [190] "NO3(L)"            "SAD(-AMMC)"        "LO(R)"            
#> [193] "FBl6"              "AVLP(R)"           "AL-DA2(R)"        
#> [196] "AL-VP5(R)"         "ICL(R)"            "AL-VC2(R)"        
#> [199] "vACA(R)"           "NO3(R)"            "AL-VA4(R)"        
#> [202] "AL-DP1m(R)"        "EBr2r4"            "PB(L6)"           
#> [205] "NO(L)"             "IB"                "g3(R)"            
#> [208] "PB(L2)"            "SMP(L)"            "b2(R)"            
#> [211] "POC"               "b'L(L)"            "EBr3am"           
#> [214] "a'3(R)"            "AL-DC1(L)"         "PLP(R)"           
#> [217] "g4(R)"             "AL-DC2(R)"         "FBl3"             
#> [220] "PB(R3)"            "IPS(R)"            "GNG"              
#> [223] "AL-DA4l(R)"        "SMP(R)"            "b'L(R)"           
#> [226] "AL-VA6(L)"         "CRE(-RUB)(L)"      "EBr5"             
#> [229] "AL-DA3(L)"         "AL-VA1v(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "BU(L)"             "FBl6"              "aL(R)"            
#>   [4] "FLA(R)"            "CRE(-RUB)(L)"      "g1(R)"            
#>   [7] "PB(L2)"            "AL-DM1(L)"         "PB(R3)"           
#>  [10] "AOT(R)"            "PB(R9)"            "AL-VA7m(R)"       
#>  [13] "PB(L8)"            "EBr3d"             "FBl5"             
#>  [16] "AL-VP3(R)"         "ICL(R)"            "AL-VA1v(R)"       
#>  [19] "SAD"               "NO3(R)"            "PB(R2)"           
#>  [22] "RUB(R)"            "ATL(L)"            "CX"               
#>  [25] "GOR(R)"            "AL-VA2(R)"         "AL-VM7v(R)"       
#>  [28] "FBl1"              "GA(R)"             "AL-DM3(L)"        
#>  [31] "a'1(R)"            "AL-VA7l(R)"        "ATL(R)"           
#>  [34] "MB(R)"             "AL-DC3(R)"         "AL-DL5(L)"        
#>  [37] "PB(R1)"            "LO(R)"             "AL-VP4(R)"        
#>  [40] "hemibrain"         "GOR(L)"            "AL-DL4(L)"        
#>  [43] "a1(R)"             "NO3(L)"            "VES(R)"           
#>  [46] "AL-DL1(R)"         "AL-DM4(R)"         "SCL(R)"           
#>  [49] "SNP(L)"            "SAD(-AMMC)"        "AL-VP2(R)"        
#>  [52] "PLP(R)"            "AL-DA4m(L)"        "AL-DP1m(L)"       
#>  [55] "WED(R)"            "EBr3pw"            "AVLP(R)"          
#>  [58] "AL-VM3(R)"         "PB(R5)"            "PRW"              
#>  [61] "ICL(L)"            "AL-VL1(R)"         "AL-VL2a(R)"       
#>  [64] "PVLP(R)"           "FBl4"              "AL-DL3(R)"        
#>  [67] "AL-DM6(L)"         "IPS(R)"            "NO"               
#>  [70] "AL-DP1m(R)"        "SCL(L)"            "a2(R)"            
#>  [73] "AL-DA3(L)"         "MB(+ACA)(R)"       "FBl2"             
#>  [76] "bL(R)"             "NO2(R)"            "AL-DM2(R)"        
#>  [79] "LAL(L)"            "SMP(L)"            "LH(R)"            
#>  [82] "RUB(L)"            "PB(R7)"            "NO1(L)"           
#>  [85] "SNP(R)"            "AL-DM5(L)"         "AL-DP1l(R)"       
#>  [88] "AL-VA6(R)"         "LOP(R)"            "MB(L)"            
#>  [91] "NO1(R)"            "AL-DC1(L)"         "NO(L)"            
#>  [94] "AL-VM1(R)"         "AL-D(R)"           "PB(L9)"           
#>  [97] "AL-DA2(R)"         "AL-DL5(R)"         "AB(R)"            
#> [100] "AL(R)"             "AL-DL2d(R)"        "AL-VC2(R)"        
#> [103] "AL-VL2p(R)"        "b'2(R)"            "EBr5"             
#> [106] "AL-DM6(R)"         "GC"                "FBl7"             
#> [109] "PB(L5)"            "INP"               "AL-DC4(R)"        
#> [112] "AL-VM5d(R)"        "AL-VM2(R)"         "PB"               
#> [115] "gL(L)"             "NO(R)"             "LAL(-GA)(R)"      
#> [118] "AL-VC5(R)"         "g2(R)"             "AL-VA6(L)"        
#> [121] "FB-column3"        "b2(R)"             "AL-VM7d(L)"       
#> [124] "AL-VC3m(R)"        "AL-DM2(L)"         "AL-DL4(R)"        
#> [127] "AL-VA4(R)"         "AL-DA3(R)"         "IB"               
#> [130] "g3(R)"             "g4(R)"             "FBl8"             
#> [133] "AL-DM4(L)"         "AL-VM7d(R)"        "PED(R)"           
#> [136] "ME(R)"             "EB"                "AMMC"             
#> [139] "dACA(R)"           "LX(L)"             "AL-VM5v(R)"       
#> [142] "b'L(R)"            "GNG"               "AL-DA1(R)"        
#> [145] "a'3(R)"            "PB(L3)"            "FBl9"             
#> [148] "SLP(R)"            "AL-VA5(R)"         "NO2(L)"           
#> [151] "AL-DC1(R)"         "AL-DC2(R)"         "vACA(R)"          
#> [154] "CRE(-ROB,-RUB)(R)" "GF(R)"             "OL(R)"            
#> [157] "VES(L)"            "mALT(L)"           "AL-VM7v(L)"       
#> [160] "bL(L)"             "AL-V(R)"           "a'L(L)"           
#> [163] "EBr1"              "PB(R8)"            "EBr2r4"           
#> [166] "SIP(L)"            "CRE(R)"            "PB(L4)"           
#> [169] "AL-VP1l(R)"        "FB"                "PB(L1)"           
#> [172] "AL-DM5(R)"         "lACA(R)"           "b1(R)"            
#> [175] "AL-VP1m(R)"        "AL-DM3(R)"         "a'L(R)"           
#> [178] "a3(R)"             "b'L(L)"            "EPA(L)"           
#> [181] "AL-VC3l(R)"        "POC"               "SPS(R)"           
#> [184] "AL-DA2(L)"         "CA(R)"             "PB(R6)"           
#> [187] "CRE(L)"            "EBr3am"            "AB(L)"            
#> [190] "VMNP"              "SIP(R)"            "CAN(R)"           
#> [193] "PENP"              "a'2(R)"            "EPA(R)"           
#> [196] "AL-VA1d(R)"        "SMP(R)"            "mALT(R)"          
#> [199] "FBl3"              "AL-DA4m(R)"        "b'1(R)"           
#> [202] "LX(R)"             "AL-DC4(L)"         "PB(R4)"           
#> [205] "BU(R)"             "LAL(R)"            "AL-VC1(R)"        
#> [208] "AL-DC2(L)"         "AME(R)"            "aL(L)"            
#> [211] "gL(R)"             "PB(L7)"            "PB(L6)"           
#> [214] "AL-DL2v(R)"        "AL-DM1(R)"         "AL-VP5(R)"        
#> [217] "AL-VP1d(R)"        "AOTU(R)"           "SPS(L)"           
#> [220] "ROB(R)"            "VLNP(R)"           "CA(L)"            
#> [223] "g5(R)"             "AL-DA4l(R)"        "AL-VA3(R)"        
#> [226] "AL-VC4(R)"         "AL-VM4(R)"         "EBr6"             
#> [229] "AL(L)"             "AL-D(L)"          
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                        
#> [1,] "POST" "/api/v:ver/npexplorer/findneurons"         
#> [2,] "GET"  "/api/v:ver/npexplorer/neuronmetavals"      
#> [3,] "GET"  "/api/raw/keyvalue/key/:instance/:key"      
#> [4,] "GET"  "/api/v:ver/raw/keyvalue/key/:instance/:key"
#> [5,] "GET"  "/api/npexplorer/autapses"                  
#> [6,] "GET"  "/api/npexplorer/completeness"              
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/raw/keyvalue.masterAPI.getKV-fm"           
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"    
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCompleteness-fm"
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }