Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $fib19
#> $fib19$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $fib19$uuid
#> [1] "6c1b926af14f43cbbcf85da97c46fb2b"
#> 
#> $fib19$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $fib19$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $fib19$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $fib19$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "dACA(R)"           "AL-DC1(R)"         "GA(R)"            
#>   [4] "g2(R)"             "AL-DM5(R)"         "a1(R)"            
#>   [7] "RUB(L)"            "AL-DA3(R)"         "AL-DA2(L)"        
#>  [10] "FB-column3"        "AL-DP1l(R)"        "AL-VC1(R)"        
#>  [13] "a'2(R)"            "BU(R)"             "AL-VC5(R)"        
#>  [16] "VMNP"              "AL-DC1(L)"         "SLP(R)"           
#>  [19] "AL-DC4(R)"         "EBr2r4"            "SAD(-AMMC)"       
#>  [22] "AL-DP1m(R)"        "EBr6"              "ATL(R)"           
#>  [25] "CA(L)"             "AL-DC4(L)"         "AL-VP1m(R)"       
#>  [28] "FBl1"              "AL-DL1(R)"         "AL-VA7l(R)"       
#>  [31] "WED(R)"            "BU(L)"             "PB"               
#>  [34] "AL-VP5(R)"         "NO1(L)"            "PB(L1)"           
#>  [37] "PB(R1)"            "g5(R)"             "LAL(L)"           
#>  [40] "AL-DA4m(L)"        "AL-DM2(R)"         "VLNP(R)"          
#>  [43] "SAD"               "AL-DC2(L)"         "IPS(R)"           
#>  [46] "AL-VA1v(R)"        "GC"                "SNP(L)"           
#>  [49] "MB(+ACA)(R)"       "FBl8"              "PB(L4)"           
#>  [52] "AL-VM4(R)"         "PB(L3)"            "AL-DM3(L)"        
#>  [55] "EPA(L)"            "AL-DM4(R)"         "AL-VM5d(R)"       
#>  [58] "AL-DP1m(L)"        "FBl2"              "a3(R)"            
#>  [61] "AL-VM3(R)"         "CRE(-ROB,-RUB)(R)" "PED(R)"           
#>  [64] "NO3(R)"            "AL-DM2(L)"         "EBr3pw"           
#>  [67] "FBl4"              "b2(R)"             "MB(L)"            
#>  [70] "bL(L)"             "AL-DL2v(R)"        "AL-VA3(R)"        
#>  [73] "AL-VA5(R)"         "EBr1"              "SIP(R)"           
#>  [76] "CAN(R)"            "AL-VL2p(R)"        "AL-VM5v(R)"       
#>  [79] "SPS(R)"            "SNP(R)"            "AL(R)"            
#>  [82] "a2(R)"             "FB"                "FBl9"             
#>  [85] "OL(R)"             "AL-DM1(R)"         "AL-VM7d(R)"       
#>  [88] "PRW"               "AL-DM4(L)"         "PB(L7)"           
#>  [91] "NO2(R)"            "AL-D(R)"           "g1(R)"            
#>  [94] "AL-VA6(L)"         "PB(R7)"            "VES(L)"           
#>  [97] "NO(L)"             "AL-VM2(R)"         "PB(L2)"           
#> [100] "VES(R)"            "ICL(R)"            "LX(L)"            
#> [103] "AL-VC3m(R)"        "AL-DA4l(R)"        "AL-VM7v(R)"       
#> [106] "AL-VP3(R)"         "lACA(R)"           "NO1(R)"           
#> [109] "AL-DM3(R)"         "ATL(L)"            "AL-DL4(L)"        
#> [112] "AL-DA1(R)"         "FBl3"              "EBr3am"           
#> [115] "AOTU(R)"           "AL-DM6(L)"         "PB(L5)"           
#> [118] "AL-DL5(L)"         "b'L(L)"            "AL-DL3(R)"        
#> [121] "PLP(R)"            "INP"               "aL(L)"            
#> [124] "AL-VA2(R)"         "MB(R)"             "a'L(R)"           
#> [127] "EPA(R)"            "AL-VM7d(L)"        "NO"               
#> [130] "PB(L6)"            "PB(R3)"            "aL(R)"            
#> [133] "SPS(L)"            "AL-VL1(R)"         "AVLP(R)"          
#> [136] "a'L(L)"            "AL-VC2(R)"         "SCL(L)"           
#> [139] "EBr5"              "PB(L8)"            "NO3(L)"           
#> [142] "AL-DC3(R)"         "AOT(R)"            "SMP(R)"           
#> [145] "g3(R)"             "hemibrain"         "AL-VA4(R)"        
#> [148] "GOR(L)"            "AL(L)"             "RUB(R)"           
#> [151] "CRE(-RUB)(L)"      "LAL(R)"            "NO2(L)"           
#> [154] "PB(R9)"            "AL-VA1d(R)"        "LOP(R)"           
#> [157] "vACA(R)"           "AMMC"              "PB(R8)"           
#> [160] "AME(R)"            "a'3(R)"            "SCL(R)"           
#> [163] "EB"                "b1(R)"             "AL-VA6(R)"        
#> [166] "AL-V(R)"           "PENP"              "b'L(R)"           
#> [169] "AL-DA3(L)"         "LH(R)"             "CRE(L)"           
#> [172] "a'1(R)"            "PVLP(R)"           "mALT(R)"          
#> [175] "AL-D(L)"           "GOR(R)"            "mALT(L)"          
#> [178] "AL-VC3l(R)"        "AL-VP4(R)"         "FBl5"             
#> [181] "AB(R)"             "AL-VL2a(R)"        "AL-VM1(R)"        
#> [184] "ME(R)"             "IB112"             "AL-DL2d(R)"       
#> [187] "AL-DM5(L)"         "PB(L9)"            "ICL(L)"           
#> [190] "AL-VA7m(R)"        "GF(R)"             "AL-DM1(L)"        
#> [193] "AL-DC2(R)"         "AL-VP1l(R)"        "AB(L)"            
#> [196] "IB"                "LO(R)"             "FBl6"             
#> [199] "gL(L)"             "AL-VC4(R)"         "EBr3d"            
#> [202] "AL-DL4(R)"         "AL-DM6(R)"         "POC"              
#> [205] "b'1(R)"            "LAL(-GA)(R)"       "LX(R)"            
#> [208] "PB(R2)"            "AL-DL5(R)"         "SMP(L)"           
#> [211] "CA(R)"             "FBl7"              "AL-VP1d(R)"       
#> [214] "b'2(R)"            "FLA(R)"            "PB(R4)"           
#> [217] "AL-DA2(R)"         "AL-DA4m(R)"        "PB(R6)"           
#> [220] "AL-VM7v(L)"        "bL(R)"             "ROB(R)"           
#> [223] "NO(R)"             "gL(R)"             "GNG"              
#> [226] "SIP(L)"            "g4(R)"             "PB(R5)"           
#> [229] "CX"                "CRE(R)"            "AL-VP2(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "AOTU(R)"           "SAD"               "PB(R7)"           
#>   [4] "SAD(-AMMC)"        "mALT(L)"           "LAL(L)"           
#>   [7] "g4(R)"             "AME(R)"            "RUB(R)"           
#>  [10] "a1(R)"             "SPS(R)"            "BU(L)"            
#>  [13] "FBl9"              "PB(L5)"            "aL(R)"            
#>  [16] "LAL(R)"            "LOP(R)"            "a'3(R)"           
#>  [19] "vACA(R)"           "FBl7"              "PB"               
#>  [22] "FBl8"              "GF(R)"             "a'L(R)"           
#>  [25] "EPA(R)"            "AOT(R)"            "EBr5"             
#>  [28] "EBr6"              "PB(L8)"            "SNP(R)"           
#>  [31] "EBr2r4"            "IPS(R)"            "SMP(L)"           
#>  [34] "PLP(R)"            "VMNP"              "RUB(L)"           
#>  [37] "PB(R5)"            "SPS(L)"            "g5(R)"            
#>  [40] "bL(R)"             "GA(R)"             "NO(R)"            
#>  [43] "FBl3"              "AB(R)"             "SIP(L)"           
#>  [46] "dACA(R)"           "AL(L)"             "a'1(R)"           
#>  [49] "VES(R)"            "NO2(R)"            "PB(R4)"           
#>  [52] "MB(R)"             "b2(R)"             "SIP(R)"           
#>  [55] "b'1(R)"            "PENP"              "MB(L)"            
#>  [58] "g2(R)"             "NO1(R)"            "CA(R)"            
#>  [61] "OL(R)"             "IB"                "SCL(L)"           
#>  [64] "CRE(R)"            "PB(L1)"            "a'2(R)"           
#>  [67] "PVLP(R)"           "FBl5"              "ATL(R)"           
#>  [70] "GOR(R)"            "b'L(L)"            "b'L(R)"           
#>  [73] "CA(L)"             "NO1(L)"            "MB(+ACA)(R)"      
#>  [76] "FBl2"              "NO(L)"             "gL(L)"            
#>  [79] "PB(L2)"            "a2(R)"             "PB(L6)"           
#>  [82] "LX(L)"             "INP"               "g3(R)"            
#>  [85] "EBr3pw"            "AL(R)"             "PB(R1)"           
#>  [88] "ICL(R)"            "PRW"               "FB"               
#>  [91] "EBr1"              "PED(R)"            "SNP(L)"           
#>  [94] "FBl4"              "PB(R3)"            "aL(L)"            
#>  [97] "VLNP(R)"           "PB(R6)"            "bL(L)"            
#> [100] "WED(R)"            "FLA(R)"            "EB"               
#> [103] "ICL(L)"            "a3(R)"             "PB(L3)"           
#> [106] "mALT(R)"           "SCL(R)"            "BU(R)"            
#> [109] "AVLP(R)"           "a'L(L)"            "hemibrain"        
#> [112] "PB(L7)"            "FBl6"              "LH(R)"            
#> [115] "VES(L)"            "EBr3d"             "LAL(-GA)(R)"      
#> [118] "GC"                "LX(R)"             "ME(R)"            
#> [121] "PB(L9)"            "GOR(L)"            "SMP(R)"           
#> [124] "CRE(-ROB,-RUB)(R)" "b1(R)"             "g1(R)"            
#> [127] "PB(R8)"            "NO"                "CAN(R)"           
#> [130] "GNG"               "PB(L4)"            "SLP(R)"           
#> [133] "lACA(R)"           "AMMC"              "NO3(L)"           
#> [136] "NO2(L)"            "PB(R9)"            "gL(R)"            
#> [139] "CRE(L)"            "FBl1"              "PB(R2)"           
#> [142] "EBr3am"            "ATL(L)"            "b'2(R)"           
#> [145] "ROB(R)"            "LO(R)"             "EPA(L)"           
#> [148] "POC"               "AB(L)"             "CX"               
#> [151] "NO3(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "SMP(L)"            "aL(L)"             "g1(R)"            
#>   [4] "AL-DM1(R)"         "EBr3d"             "IB"               
#>   [7] "AL-DC1(L)"         "CRE(-RUB)(L)"      "SAD(-AMMC)"       
#>  [10] "AL-DA3(R)"         "AL-VP1l(R)"        "AL-DA3(L)"        
#>  [13] "ICL(L)"            "FBl6"              "ICL(R)"           
#>  [16] "VLNP(R)"           "SAD"               "aL(R)"            
#>  [19] "AL-DM1(L)"         "LO(R)"             "b'L(R)"           
#>  [22] "g2(R)"             "CRE(L)"            "EBr2r4"           
#>  [25] "NO"                "NO(L)"             "AL-VC1(R)"        
#>  [28] "AL-DA4l(R)"        "AL-VC3l(R)"        "PB(R6)"           
#>  [31] "GF(R)"             "WED(R)"            "POC"              
#>  [34] "SNP(R)"            "CA(L)"             "SCL(R)"           
#>  [37] "GOR(R)"            "a3(R)"             "AL(L)"            
#>  [40] "AL-D(L)"           "AL-DM3(L)"         "AL-DP1l(R)"       
#>  [43] "FB"                "PB"                "AL-DC4(L)"        
#>  [46] "AL-DL5(L)"         "SMP(R)"            "AL-DC1(R)"        
#>  [49] "AL-DM6(R)"         "EBr5"              "EB"               
#>  [52] "PVLP(R)"           "PB(R5)"            "EBr1"             
#>  [55] "VES(R)"            "CA(R)"             "OL(R)"            
#>  [58] "AL-DP1m(R)"        "EBr6"              "AL-VA6(L)"        
#>  [61] "AME(R)"            "NO3(R)"            "PB(R3)"           
#>  [64] "mALT(L)"           "FBl5"              "VES(L)"           
#>  [67] "a2(R)"             "mALT(R)"           "a'L(R)"           
#>  [70] "AL-DM6(L)"         "AL-VM7d(L)"        "PB(R1)"           
#>  [73] "PB(L7)"            "AL-DA4m(R)"        "PB(L2)"           
#>  [76] "LOP(R)"            "g3(R)"             "FBl1"             
#>  [79] "RUB(R)"            "a'3(R)"            "vACA(R)"          
#>  [82] "AOT(R)"            "PENP"              "dACA(R)"          
#>  [85] "NO2(R)"            "AL-VC5(R)"         "AL-VP2(R)"        
#>  [88] "AL-VP1m(R)"        "PRW"               "AL-DP1m(L)"       
#>  [91] "b1(R)"             "bL(R)"             "AMMC"             
#>  [94] "lACA(R)"           "AL-DA4m(L)"        "EBr3am"           
#>  [97] "AL-VA4(R)"         "AL-VM7d(R)"        "LAL(L)"           
#> [100] "LX(L)"             "b'L(L)"            "LAL(R)"           
#> [103] "AL-DM3(R)"         "AL-VA1v(R)"        "GA(R)"            
#> [106] "LAL(-GA)(R)"       "GC"                "ROB(R)"           
#> [109] "IPS(R)"            "AL-V(R)"           "AL-DM5(L)"        
#> [112] "CAN(R)"            "AL-DL2v(R)"        "AL-VM2(R)"        
#> [115] "PB(L5)"            "AL-VM4(R)"         "SPS(R)"           
#> [118] "AL-DM4(L)"         "PB(R4)"            "AL-VM1(R)"        
#> [121] "AL-VM5d(R)"        "PB(L1)"            "bL(L)"            
#> [124] "FBl7"              "RUB(L)"            "AL-DA2(R)"        
#> [127] "AL-VP1d(R)"        "NO(R)"             "AL-VA1d(R)"       
#> [130] "AL-DL1(R)"         "AL-D(R)"           "b2(R)"            
#> [133] "FLA(R)"            "AL-DC2(L)"         "AL-DL4(L)"        
#> [136] "PB(L9)"            "FBl3"              "AL-VP3(R)"        
#> [139] "PB(L3)"            "hemibrain"         "BU(L)"            
#> [142] "AL-DM4(R)"         "EPA(R)"            "AL-DM2(L)"        
#> [145] "SIP(R)"            "AL-DL3(R)"         "AL-VM3(R)"        
#> [148] "AL(R)"             "AL-VL1(R)"         "AL-VL2p(R)"       
#> [151] "PB(R7)"            "VMNP"              "PB(L8)"           
#> [154] "PB(L6)"            "ATL(L)"            "GOR(L)"           
#> [157] "MB(L)"             "AL-DC4(R)"         "AL-VA6(R)"        
#> [160] "AL-VA7l(R)"        "AL-VC3m(R)"        "AOTU(R)"          
#> [163] "FBl9"              "AL-VA5(R)"         "AL-VL2a(R)"       
#> [166] "AL-VM5v(R)"        "SCL(L)"            "LX(R)"            
#> [169] "PB(L4)"            "a'L(L)"            "AL-DC3(R)"        
#> [172] "SLP(R)"            "EBr3pw"            "SNP(L)"           
#> [175] "INP"               "AL-DL5(R)"         "SIP(L)"           
#> [178] "CRE(R)"            "ME(R)"             "gL(L)"            
#> [181] "a'2(R)"            "NO3(L)"            "FB-column3"       
#> [184] "AB(R)"             "AL-DL4(R)"         "AL-DM2(R)"        
#> [187] "AL-VP4(R)"         "a1(R)"             "AVLP(R)"          
#> [190] "AL-VM7v(R)"        "AL-DA2(L)"         "AL-DC2(R)"        
#> [193] "GNG"               "SPS(L)"            "PLP(R)"           
#> [196] "a'1(R)"            "MB(+ACA)(R)"       "ATL(R)"           
#> [199] "AL-VA7m(R)"        "CRE(-ROB,-RUB)(R)" "g4(R)"            
#> [202] "AL-VA2(R)"         "AL-DA1(R)"         "AL-VA3(R)"        
#> [205] "AL-VC2(R)"         "AB(L)"             "FBl8"             
#> [208] "MB(R)"             "NO2(L)"            "NO1(L)"           
#> [211] "FBl2"              "AL-VC4(R)"         "AL-VP5(R)"        
#> [214] "CX"                "AL-DM5(R)"         "b'2(R)"           
#> [217] "PED(R)"            "BU(R)"             "AL-VM7v(L)"       
#> [220] "FBl4"              "NO1(R)"            "PB(R9)"           
#> [223] "gL(R)"             "LH(R)"             "EPA(L)"           
#> [226] "AL-DL2d(R)"        "PB(R8)"            "g5(R)"            
#> [229] "PB(R2)"            "b'1(R)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "EBr3d"             "GNG"               "AL-VC4(R)"        
#>   [4] "ATL(L)"            "ME(R)"             "FBl4"             
#>   [7] "AL-DL3(R)"         "IPS(R)"            "AL-DC2(R)"        
#>  [10] "PB(L6)"            "bL(L)"             "AL-DA2(R)"        
#>  [13] "POC"               "b'2(R)"            "PED(R)"           
#>  [16] "g4(R)"             "FBl5"              "FBl6"             
#>  [19] "aL(L)"             "g3(R)"             "BU(L)"            
#>  [22] "bL(R)"             "PB(R6)"            "VMNP"             
#>  [25] "AL-VM3(R)"         "FB"                "AL-DM3(R)"        
#>  [28] "ATL(R)"            "AL-VP1m(R)"        "g1(R)"            
#>  [31] "a'1(R)"            "AL-DL5(L)"         "MB(+ACA)(R)"      
#>  [34] "AL-DL2v(R)"        "GA(R)"             "AL-DC1(L)"        
#>  [37] "PB(L4)"            "MB(L)"             "CRE(L)"           
#>  [40] "ROB(R)"            "NO(R)"             "NO3(L)"           
#>  [43] "FB-column3"        "AL-VL2p(R)"        "LX(R)"            
#>  [46] "AL-DC4(L)"         "PB(R2)"            "AL-VM5v(R)"       
#>  [49] "NO"                "FBl7"              "CAN(R)"           
#>  [52] "NO1(R)"            "AL-DM5(L)"         "g5(R)"            
#>  [55] "SLP(R)"            "AL-DM1(R)"         "AL-VL1(R)"        
#>  [58] "g2(R)"             "AL(L)"             "AL-DM6(R)"        
#>  [61] "AL-VC3l(R)"        "PB(L7)"            "MB(R)"            
#>  [64] "AL-VP1l(R)"        "PB(L1)"            "SNP(R)"           
#>  [67] "AL-DM4(R)"         "PB(L2)"            "SAD"              
#>  [70] "CRE(R)"            "AL-VM4(R)"         "VES(L)"           
#>  [73] "RUB(R)"            "GC"                "EBr2r4"           
#>  [76] "FBl2"              "ICL(R)"            "AL-DL1(R)"        
#>  [79] "AL-D(R)"           "AL-VA7l(R)"        "EBr3am"           
#>  [82] "dACA(R)"           "AL-DA3(L)"         "CA(L)"            
#>  [85] "AL-DM6(L)"         "SIP(R)"            "AL-VA3(R)"        
#>  [88] "AL-VA7m(R)"        "LAL(-GA)(R)"       "AL-DA2(L)"        
#>  [91] "AL-DA4m(R)"        "AL-DL4(R)"         "CX"               
#>  [94] "PB(L3)"            "AL-DA1(R)"         "LH(R)"            
#>  [97] "b'L(L)"            "AL-DA4l(R)"        "AL-DM3(L)"        
#> [100] "EBr3pw"            "RUB(L)"            "a1(R)"            
#> [103] "a2(R)"             "EB"                "PB(L8)"           
#> [106] "AL-VM2(R)"         "EPA(L)"            "NO2(L)"           
#> [109] "AL-DM2(L)"         "SAD(-AMMC)"        "AL-DM1(L)"        
#> [112] "AL-VM7v(L)"        "CRE(-RUB)(L)"      "NO3(R)"           
#> [115] "EBr1"              "b1(R)"             "INP"              
#> [118] "PB(R4)"            "AL-VM5d(R)"        "LO(R)"            
#> [121] "FBl8"              "AL-VP2(R)"         "PB(R3)"           
#> [124] "AL-VP5(R)"         "EPA(R)"            "AOTU(R)"          
#> [127] "lACA(R)"           "PB"                "BU(R)"            
#> [130] "CA(R)"             "AB(R)"             "AL-DP1l(R)"       
#> [133] "AL-VA5(R)"         "NO1(L)"            "SCL(R)"           
#> [136] "AVLP(R)"           "AL-VC2(R)"         "AL-VC5(R)"        
#> [139] "AL-VM7v(R)"        "a'2(R)"            "PVLP(R)"          
#> [142] "SPS(R)"            "gL(L)"             "a3(R)"            
#> [145] "AL-VP1d(R)"        "mALT(L)"           "EBr6"             
#> [148] "AL-V(R)"           "AL-DM4(L)"         "AMMC"             
#> [151] "SMP(L)"            "AL-D(L)"           "vACA(R)"          
#> [154] "PB(R1)"            "FBl3"              "SCL(L)"           
#> [157] "CRE(-ROB,-RUB)(R)" "PRW"               "AL-VA2(R)"        
#> [160] "AL-VC1(R)"         "SMP(R)"            "a'3(R)"           
#> [163] "AL-DL4(L)"         "a'L(L)"            "PLP(R)"           
#> [166] "AL-VM1(R)"         "ICL(L)"            "LAL(R)"           
#> [169] "AL-DC4(R)"         "hemibrain"         "AL-DC1(R)"        
#> [172] "AL-VA6(R)"         "AB(L)"             "LAL(L)"           
#> [175] "SPS(L)"            "AL-DA3(R)"         "AL-DL5(R)"        
#> [178] "AL-VM7d(L)"        "FBl1"              "AL-DM5(R)"        
#> [181] "AOT(R)"            "mALT(R)"           "AL-DC2(L)"        
#> [184] "PB(R7)"            "PB(R5)"            "AL-VM7d(R)"       
#> [187] "IB"                "AL(R)"             "AL-VP3(R)"        
#> [190] "PENP"              "GOR(L)"            "PB(R8)"           
#> [193] "VES(R)"            "NO2(R)"            "AL-VL2a(R)"       
#> [196] "VLNP(R)"           "b2(R)"             "b'L(R)"           
#> [199] "LX(L)"             "WED(R)"            "SNP(L)"           
#> [202] "AL-DL2d(R)"        "AL-VA4(R)"         "b'1(R)"           
#> [205] "LOP(R)"            "AL-VA6(L)"         "FBl9"             
#> [208] "AL-VC3m(R)"        "gL(R)"             "PB(L9)"           
#> [211] "AL-DC3(R)"         "AL-DM2(R)"         "aL(R)"            
#> [214] "EBr5"              "AL-VA1d(R)"        "FLA(R)"           
#> [217] "AME(R)"            "AL-DP1m(L)"        "GOR(R)"           
#> [220] "a'L(R)"            "AL-DA4m(L)"        "NO(L)"            
#> [223] "AL-VA1v(R)"        "AL-VP4(R)"         "GF(R)"            
#> [226] "OL(R)"             "SIP(L)"            "PB(L5)"           
#> [229] "PB(R9)"            "AL-DP1m(R)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                           
#> [1,] "GET"  "/api"                                         
#> [2,] "GET"  "/api/dbmeta/database"                         
#> [3,] "GET"  "/api/v:ver/dbmeta/database"                   
#> [4,] "GET"  "/api/npexplorer/neuronmeta"                   
#> [5,] "POST" "/api/v:ver/npexplorer/roiconnectivity"        
#> [6,] "GET"  "/api/v:ver/npexplorer/celltype/:dataset/:type"
#>      name                                                                                      
#> [1,] "github.com/labstack/echo.glob..func1"                                                    
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatabase-fm"          
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                      
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMeta-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                      
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                      
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }