Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "SPS(R)"            "mALT(R)"           "AL(L)"            
#>   [4] "AB(R)"             "AOT(R)"            "a2(R)"            
#>   [7] "PB(R8)"            "AME(R)"            "NO2(R)"           
#>  [10] "LOP(R)"            "PB(L5)"            "g3(R)"            
#>  [13] "LAL(-GA)(R)"       "b'L(R)"            "EB"               
#>  [16] "aL(L)"             "hemibrain"         "NO(R)"            
#>  [19] "ATL(R)"            "PB(L4)"            "vACA(R)"          
#>  [22] "gL(R)"             "g5(R)"             "LO(R)"            
#>  [25] "VES(R)"            "SCL(L)"            "b2(R)"            
#>  [28] "EBr5"              "b1(R)"             "PB(R4)"           
#>  [31] "a'3(R)"            "LH(R)"             "AOTU(R)"          
#>  [34] "GC"                "NO1(R)"            "SPS(L)"           
#>  [37] "LX(R)"             "PB(L6)"            "EBr3pw"           
#>  [40] "OL(R)"             "PB(R6)"            "b'2(R)"           
#>  [43] "CRE(L)"            "SAD(-AMMC)"        "AL(R)"            
#>  [46] "BU(R)"             "NO(L)"             "MB(R)"            
#>  [49] "POC"               "mALT(L)"           "FBl1"             
#>  [52] "b'1(R)"            "GA(R)"             "BU(L)"            
#>  [55] "EPA(L)"            "bL(R)"             "SLP(R)"           
#>  [58] "PENP"              "PED(R)"            "g2(R)"            
#>  [61] "PB(R1)"            "PB(R3)"            "SAD"              
#>  [64] "PVLP(R)"           "FBl4"              "FBl5"             
#>  [67] "PB(L2)"            "NO2(L)"            "lACA(R)"          
#>  [70] "PB(R7)"            "FB"                "PB"               
#>  [73] "g4(R)"             "MB(+ACA)(R)"       "RUB(R)"           
#>  [76] "NO3(R)"            "EBr2r4"            "NO"               
#>  [79] "PRW"               "gL(L)"             "SNP(L)"           
#>  [82] "PB(R2)"            "LAL(R)"            "GOR(L)"           
#>  [85] "ME(R)"             "ICL(R)"            "GOR(R)"           
#>  [88] "a'L(L)"            "ATL(L)"            "FBl3"             
#>  [91] "FBl9"              "INP"               "a'2(R)"           
#>  [94] "IPS(R)"            "LX(L)"             "VMNP"             
#>  [97] "PB(R9)"            "CA(R)"             "CAN(R)"           
#> [100] "AMMC"              "ICL(L)"            "a3(R)"            
#> [103] "CX"                "PB(L3)"            "SCL(R)"           
#> [106] "dACA(R)"           "a'1(R)"            "bL(L)"            
#> [109] "a1(R)"             "GF(R)"             "SIP(L)"           
#> [112] "FLA(R)"            "ROB(R)"            "b'L(L)"           
#> [115] "WED(R)"            "EBr6"              "AVLP(R)"          
#> [118] "GNG"               "EBr3am"            "CRE(-ROB,-RUB)(R)"
#> [121] "AB(L)"             "SMP(L)"            "NO1(L)"           
#> [124] "CRE(R)"            "EBr1"              "PB(L8)"           
#> [127] "EPA(R)"            "VLNP(R)"           "aL(R)"            
#> [130] "g1(R)"             "VES(L)"            "CA(L)"            
#> [133] "PB(L1)"            "PLP(R)"            "EBr3d"            
#> [136] "SIP(R)"            "FBl2"              "FBl6"             
#> [139] "MB(L)"             "SNP(R)"            "NO3(L)"           
#> [142] "PB(L7)"            "IB"                "LAL(L)"           
#> [145] "a'L(R)"            "FBl7"              "FBl8"             
#> [148] "SMP(R)"            "PB(L9)"            "RUB(L)"           
#> [151] "PB(R5)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method  path                                  
#> [1,] "GET"   "/login"                              
#> [2,] "POST"  "/api/npexplorer/roiconnectivity"     
#> [3,] "TRACE" "/api"                                
#> [4,] "GET"   "/api/npexplorer/distribution"        
#> [5,] "GET"   "/api/v:ver/cached/roicompleteness"   
#> [6,] "GET"   "/api/v:ver/npexplorer/neuronmetavals"
#>      name                                                                                           
#> [1,] "github.com/janelia-flyem/echo-secure.loginHandler"                                            
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIConnectivity-fm"
#> [3,] "github.com/labstack/echo.glob..func1"                                                         
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"   
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }