Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(R6)"            "NO3(R)"            "NO1(R)"           
#>   [4] "bL(R)"             "AL(R)"             "SCL(R)"           
#>   [7] "FBl3"              "LH(R)"             "a3(R)"            
#>  [10] "CRE(R)"            "VMNP"              "FBl7"             
#>  [13] "AME(R)"            "AB(L)"             "PB(R2)"           
#>  [16] "IPS(R)"            "NO(L)"             "b'2(R)"           
#>  [19] "g3(R)"             "SIP(L)"            "dACA(R)"          
#>  [22] "EBr3d"             "AOT(R)"            "GF(R)"            
#>  [25] "NO(R)"             "LAL(L)"            "bL(L)"            
#>  [28] "OL(R)"             "SMP(R)"            "a'3(R)"           
#>  [31] "AOTU(R)"           "GC"                "MB(L)"            
#>  [34] "MB(+ACA)(R)"       "vACA(R)"           "NO1(L)"           
#>  [37] "b'1(R)"            "ATL(L)"            "PENP"             
#>  [40] "mALT(R)"           "a'L(R)"            "PB(L9)"           
#>  [43] "FBl8"              "PB(L1)"            "SIP(R)"           
#>  [46] "PB(R9)"            "ICL(R)"            "a1(R)"            
#>  [49] "FBl6"              "LAL(R)"            "SLP(R)"           
#>  [52] "SMP(L)"            "SAD"               "BU(L)"            
#>  [55] "PB(L7)"            "SAD(-AMMC)"        "SCL(L)"           
#>  [58] "AMMC"              "EB"                "FBl5"             
#>  [61] "EPA(L)"            "b'L(L)"            "PB"               
#>  [64] "gL(L)"             "g4(R)"             "ROB(R)"           
#>  [67] "AVLP(R)"           "EBr6"              "PB(L6)"           
#>  [70] "PB(L3)"            "a2(R)"             "aL(R)"            
#>  [73] "PB(R8)"            "BU(R)"             "PB(R7)"           
#>  [76] "b'L(R)"            "VES(R)"            "b1(R)"            
#>  [79] "GNG"               "EBr1"              "ICL(L)"           
#>  [82] "FBl4"              "IB"                "a'2(R)"           
#>  [85] "EBr3pw"            "INP"               "POC"              
#>  [88] "RUB(R)"            "VLNP(R)"           "EBr5"             
#>  [91] "PB(L8)"            "CRE(L)"            "AB(R)"            
#>  [94] "NO2(R)"            "PED(R)"            "a'L(L)"           
#>  [97] "EBr3am"            "PLP(R)"            "ME(R)"            
#> [100] "SNP(L)"            "PB(L5)"            "CRE(-ROB,-RUB)(R)"
#> [103] "FBl1"              "CAN(R)"            "GOR(R)"           
#> [106] "mALT(L)"           "LAL(-GA)(R)"       "b2(R)"            
#> [109] "SPS(L)"            "g5(R)"             "PVLP(R)"          
#> [112] "CX"                "LX(L)"             "RUB(L)"           
#> [115] "PB(R4)"            "hemibrain"         "g1(R)"            
#> [118] "EPA(R)"            "a'1(R)"            "NO3(L)"           
#> [121] "PB(R5)"            "GOR(L)"            "g2(R)"            
#> [124] "SNP(R)"            "VES(L)"            "WED(R)"           
#> [127] "gL(R)"             "FBl2"              "AL(L)"            
#> [130] "NO"                "PB(R3)"            "SPS(R)"           
#> [133] "lACA(R)"           "FLA(R)"            "LX(R)"            
#> [136] "EBr2r4"            "MB(R)"             "PB(L4)"           
#> [139] "PB(L2)"            "CA(L)"             "PRW"              
#> [142] "PB(R1)"            "FBl9"              "CA(R)"            
#> [145] "aL(L)"             "NO2(L)"            "GA(R)"            
#> [148] "FB"                "LO(R)"             "ATL(R)"           
#> [151] "LOP(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                                   
#> [1,] "CONNECT" "/api/*"                               
#> [2,] "GET"     "/api/v:ver/cached/roiconnectivity"    
#> [3,] "GET"     "/api/npexplorer/neuronmeta"           
#> [4,] "POST"    "/api/v:ver/npexplorer/roiconnectivity"
#> [5,] "GET"     "/api/skeletons/skeleton/:dataset/:id" 
#> [6,] "PATCH"   "/api"                                 
#>      name                                                                                      
#> [1,] "github.com/labstack/echo.glob..func1"                                                    
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                      
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMeta-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                      
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/skeletons.masterAPI.getSkeleton-fm"      
#> [6,] "github.com/labstack/echo.glob..func1"                                                    

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }