Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(L3)"            "mALT(R)"           "PB(R5)"           
#>   [4] "ATL(L)"            "NO"                "EBr3pw"           
#>   [7] "RUB(L)"            "SCL(L)"            "mALT(L)"          
#>  [10] "b'2(R)"            "LAL(-GA)(R)"       "SCL(R)"           
#>  [13] "EBr5"              "NO1(R)"            "EB"               
#>  [16] "ICL(L)"            "LO(R)"             "PB(R2)"           
#>  [19] "MB(L)"             "PB(L7)"            "SNP(L)"           
#>  [22] "FBl7"              "bL(R)"             "PED(R)"           
#>  [25] "g2(R)"             "dACA(R)"           "SNP(R)"           
#>  [28] "a1(R)"             "lACA(R)"           "AL(L)"            
#>  [31] "PVLP(R)"           "ATL(R)"            "PENP"             
#>  [34] "PB(L8)"            "PLP(R)"            "FBl2"             
#>  [37] "FBl9"              "NO(L)"             "EBr3am"           
#>  [40] "PB(L9)"            "CRE(L)"            "FBl6"             
#>  [43] "SLP(R)"            "a2(R)"             "ROB(R)"           
#>  [46] "PB(R1)"            "NO3(L)"            "MB(R)"            
#>  [49] "vACA(R)"           "GOR(L)"            "aL(R)"            
#>  [52] "a'L(R)"            "IPS(R)"            "NO2(L)"           
#>  [55] "LAL(L)"            "BU(L)"             "MB(+ACA)(R)"      
#>  [58] "FB"                "gL(R)"             "g1(R)"            
#>  [61] "g4(R)"             "a'2(R)"            "AOT(R)"           
#>  [64] "CRE(-ROB,-RUB)(R)" "AMMC"              "PB(L1)"           
#>  [67] "NO(R)"             "FBl5"              "GOR(R)"           
#>  [70] "a'L(L)"            "RUB(R)"            "PB(R9)"           
#>  [73] "GC"                "NO3(R)"            "gL(L)"            
#>  [76] "g5(R)"             "INP"               "VES(L)"           
#>  [79] "SMP(R)"            "SPS(R)"            "a3(R)"            
#>  [82] "AL(R)"             "CX"                "SPS(L)"           
#>  [85] "FBl3"              "PB(R4)"            "OL(R)"            
#>  [88] "CA(L)"             "SAD"               "ME(R)"            
#>  [91] "SIP(R)"            "a'1(R)"            "AB(R)"            
#>  [94] "EPA(L)"            "LOP(R)"            "PB(R6)"           
#>  [97] "b'L(R)"            "FLA(R)"            "SAD(-AMMC)"       
#> [100] "NO2(R)"            "LX(L)"             "EBr6"             
#> [103] "PB(R8)"            "AME(R)"            "CAN(R)"           
#> [106] "SIP(L)"            "CRE(R)"            "CA(R)"            
#> [109] "LAL(R)"            "AOTU(R)"           "PB(R7)"           
#> [112] "aL(L)"             "PRW"               "WED(R)"           
#> [115] "b2(R)"             "VMNP"              "EBr2r4"           
#> [118] "PB(L6)"            "bL(L)"             "FBl4"             
#> [121] "g3(R)"             "NO1(L)"            "POC"              
#> [124] "IB"                "a'3(R)"            "GF(R)"            
#> [127] "GA(R)"             "LX(R)"             "LH(R)"            
#> [130] "b'L(L)"            "SMP(L)"            "PB(R3)"           
#> [133] "AVLP(R)"           "b1(R)"             "hemibrain"        
#> [136] "EBr1"              "EPA(R)"            "VES(R)"           
#> [139] "FBl1"              "ICL(R)"            "GNG"              
#> [142] "PB(L5)"            "BU(R)"             "PB(L4)"           
#> [145] "AB(L)"             "PB(L2)"            "PB"               
#> [148] "FBl8"              "EBr3d"             "VLNP(R)"          
#> [151] "b'1(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "gL(L)"             "AL-DA4m(L)"        "AL-DM1(L)"        
#>   [4] "aL(L)"             "SNP(R)"            "FBl7"             
#>   [7] "hemibrain"         "lACA(R)"           "EBr1"             
#>  [10] "gL(R)"             "PB(R2)"            "CX"               
#>  [13] "GF(R)"             "GC"                "AL-D(L)"          
#>  [16] "AL-DM3(L)"         "AL-DM5(L)"         "a2(R)"            
#>  [19] "AL-DL2d(R)"        "FBl2"              "aL(R)"            
#>  [22] "AL-VC3l(R)"        "g1(R)"             "mALT(L)"          
#>  [25] "AL-DL1(R)"         "g2(R)"             "EPA(R)"           
#>  [28] "PVLP(R)"           "FBl4"              "AME(R)"           
#>  [31] "AL-VM5v(R)"        "vACA(R)"           "mALT(R)"          
#>  [34] "AL-DP1m(R)"        "AL-VA2(R)"         "LAL(L)"           
#>  [37] "AL-DC1(R)"         "AL-VC3m(R)"        "EBr6"             
#>  [40] "a'2(R)"            "FBl9"              "CAN(R)"           
#>  [43] "GOR(R)"            "CA(L)"             "AOTU(R)"          
#>  [46] "b'L(L)"            "ME(R)"             "AL-V(R)"          
#>  [49] "POC"               "AL-DL3(R)"         "PB(L8)"           
#>  [52] "PB"                "AL-DP1m(L)"        "EBr3pw"           
#>  [55] "AL-DA2(R)"         "AL-DL5(R)"         "SPS(R)"           
#>  [58] "PB(R6)"            "g4(R)"             "SNP(L)"           
#>  [61] "LX(R)"             "a'L(R)"            "FB"               
#>  [64] "AL-DA4m(R)"        "AL-VA1v(R)"        "IB"               
#>  [67] "EBr3d"             "SMP(R)"            "AL-VA5(R)"        
#>  [70] "b1(R)"             "CRE(-ROB,-RUB)(R)" "b'2(R)"           
#>  [73] "AL-VL2p(R)"        "VLNP(R)"           "SCL(L)"           
#>  [76] "OL(R)"             "AL-DP1l(R)"        "AL-VC5(R)"        
#>  [79] "AL-VP2(R)"         "AMMC"              "AL-DC3(R)"        
#>  [82] "AL-D(R)"           "AL-DA4l(R)"        "PB(R7)"           
#>  [85] "BU(R)"             "AVLP(R)"           "AL-VP4(R)"        
#>  [88] "SAD"               "VMNP"              "ICL(R)"           
#>  [91] "AB(R)"             "AL-DC2(R)"         "SCL(R)"           
#>  [94] "FBl3"              "LX(L)"             "GNG"              
#>  [97] "PB(L4)"            "VES(L)"            "AL-DL4(R)"        
#> [100] "AL-DM6(R)"         "PB(L7)"            "PB(R5)"           
#> [103] "AL-DC1(L)"         "PB(R4)"            "EBr3am"           
#> [106] "FBl1"              "AL-VL1(R)"         "PB(R9)"           
#> [109] "FBl5"              "FB-column3"        "NO"               
#> [112] "ATL(R)"            "PB(L2)"            "NO1(R)"           
#> [115] "FLA(R)"            "AL-DC2(L)"         "NO(R)"            
#> [118] "LO(R)"             "CA(R)"             "AL-DM3(R)"        
#> [121] "EBr5"              "MB(R)"             "AL-VM1(R)"        
#> [124] "PB(L5)"            "AL-DM4(L)"         "AL-VL2a(R)"       
#> [127] "AL-VC2(R)"         "ICL(L)"            "AL-DM1(R)"        
#> [130] "AL-DM4(R)"         "CRE(R)"            "AL-VM7d(L)"       
#> [133] "bL(L)"             "AL-VP3(R)"         "GA(R)"            
#> [136] "CRE(-RUB)(L)"      "RUB(R)"            "b'1(R)"           
#> [139] "PB(L9)"            "AL-VC1(R)"         "AL-VP1l(R)"       
#> [142] "LAL(R)"            "WED(R)"            "MB(+ACA)(R)"      
#> [145] "a3(R)"             "AL-DA1(R)"         "PRW"              
#> [148] "AL-DA3(L)"         "ROB(R)"            "LH(R)"            
#> [151] "bL(R)"             "PB(L3)"            "PB(R8)"           
#> [154] "CRE(L)"            "FBl8"              "PB(R3)"           
#> [157] "AL-DA3(R)"         "GOR(L)"            "SIP(L)"           
#> [160] "PED(R)"            "AL-DC4(R)"         "PENP"             
#> [163] "LAL(-GA)(R)"       "EBr2r4"            "AL-VA6(L)"        
#> [166] "NO(L)"             "a'L(L)"            "AL(R)"            
#> [169] "AL-VA1d(R)"        "g3(R)"             "AL-DM6(L)"        
#> [172] "EPA(L)"            "AL-VA3(R)"         "AL-VA6(R)"        
#> [175] "PB(L1)"            "RUB(L)"            "AL-VP1d(R)"       
#> [178] "dACA(R)"           "SIP(R)"            "IPS(R)"           
#> [181] "LOP(R)"            "EB"                "a'3(R)"           
#> [184] "AL-VA7l(R)"        "AL-VM7v(R)"        "AL-VM5d(R)"       
#> [187] "PB(R1)"            "SPS(L)"            "SAD(-AMMC)"       
#> [190] "INP"               "NO1(L)"            "AL-VM7d(R)"       
#> [193] "ATL(L)"            "NO3(R)"            "AL-DL4(L)"        
#> [196] "FBl6"              "a1(R)"             "NO2(L)"           
#> [199] "PLP(R)"            "BU(L)"             "AOT(R)"           
#> [202] "MB(L)"             "AL-DM2(R)"         "b'L(R)"           
#> [205] "AL-VM3(R)"         "AL-VC4(R)"         "NO3(L)"           
#> [208] "AL-DL2v(R)"        "AL-VM2(R)"         "AL-VP5(R)"        
#> [211] "AL-VP1m(R)"        "AL-DC4(L)"         "g5(R)"            
#> [214] "AL-DM2(L)"         "AL(L)"             "AL-VM7v(L)"       
#> [217] "PB(L6)"            "AL-VA7m(R)"        "SMP(L)"           
#> [220] "AL-DL5(L)"         "a'1(R)"            "VES(R)"           
#> [223] "SLP(R)"            "AL-VM4(R)"         "AL-DA2(L)"        
#> [226] "NO2(R)"            "b2(R)"             "AL-VA4(R)"        
#> [229] "AB(L)"             "AL-DM5(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                   
#> [1,] "POST" "/api/v:ver/npexplorer/neuronmetavals" 
#> [2,] "GET"  "/api/v:ver/npexplorer/roiconnectivity"
#> [3,] "POST" "/api/npexplorer/simpleconnections"    
#> [4,] "POST" "/api/v:ver/npexplorer/roisinneuron"   
#> [5,] "POST" "/logout"                              
#> [6,] "GET"  "/api"                                 
#>      name                                                                                             
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [5,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                         
#> [6,] "github.com/labstack/echo.glob..func1"                                                           

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }