Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(R6)"            "SPS(R)"            "EBr1"             
#>   [4] "a'1(R)"            "PB(R7)"            "mALT(R)"          
#>   [7] "NO3(R)"            "NO3(L)"            "POC"              
#>  [10] "AOTU(R)"           "SNP(R)"            "FBl6"             
#>  [13] "FBl7"              "SMP(L)"            "g2(R)"            
#>  [16] "NO"                "CA(L)"             "SIP(R)"           
#>  [19] "PB(R9)"            "NO(L)"             "IB"               
#>  [22] "INP"               "MB(L)"             "dACA(R)"          
#>  [25] "SPS(L)"            "EBr2r4"            "g4(R)"            
#>  [28] "PB(L6)"            "VMNP"              "NO(R)"            
#>  [31] "FBl9"              "mALT(L)"           "BU(L)"            
#>  [34] "PB(L1)"            "BU(R)"             "CRE(L)"           
#>  [37] "b'L(L)"            "PB(L5)"            "EPA(L)"           
#>  [40] "PED(R)"            "ME(R)"             "EB"               
#>  [43] "AL(L)"             "EBr3pw"            "PVLP(R)"          
#>  [46] "OL(R)"             "LAL(L)"            "SAD"              
#>  [49] "GC"                "VES(R)"            "FBl3"             
#>  [52] "ATL(R)"            "SIP(L)"            "EBr6"             
#>  [55] "PB(R8)"            "LX(R)"             "g5(R)"            
#>  [58] "WED(R)"            "b2(R)"             "FB"               
#>  [61] "PB"                "FBl4"              "lACA(R)"          
#>  [64] "g1(R)"             "aL(R)"             "EBr3am"           
#>  [67] "b'2(R)"            "g3(R)"             "PB(R1)"           
#>  [70] "ICL(R)"            "NO2(L)"            "b1(R)"            
#>  [73] "SLP(R)"            "SCL(L)"            "a1(R)"            
#>  [76] "GA(R)"             "LAL(-GA)(R)"       "LO(R)"            
#>  [79] "FBl5"              "SCL(R)"            "a'2(R)"           
#>  [82] "gL(R)"             "PB(L9)"            "hemibrain"        
#>  [85] "CRE(R)"            "CA(R)"             "AVLP(R)"          
#>  [88] "PB(R3)"            "b'L(R)"            "aL(L)"            
#>  [91] "bL(R)"             "PB(L4)"            "CRE(-ROB,-RUB)(R)"
#>  [94] "EBr3d"             "ROB(R)"            "AB(R)"            
#>  [97] "GOR(R)"            "FBl8"              "LAL(R)"           
#> [100] "VLNP(R)"           "SNP(L)"            "FBl1"             
#> [103] "RUB(L)"            "ATL(L)"            "SMP(R)"           
#> [106] "b'1(R)"            "a'L(L)"            "PB(R5)"           
#> [109] "GOR(L)"            "ICL(L)"            "PB(L7)"           
#> [112] "PB(R2)"            "CAN(R)"            "GNG"              
#> [115] "NO1(L)"            "EPA(R)"            "NO2(R)"           
#> [118] "LX(L)"             "AL(R)"             "AOT(R)"           
#> [121] "PB(L8)"            "NO1(R)"            "a'L(R)"           
#> [124] "MB(+ACA)(R)"       "a'3(R)"            "CX"               
#> [127] "PB(L3)"            "VES(L)"            "MB(R)"            
#> [130] "AB(L)"             "PB(L2)"            "PLP(R)"           
#> [133] "LH(R)"             "IPS(R)"            "PENP"             
#> [136] "GF(R)"             "EBr5"              "FLA(R)"           
#> [139] "gL(L)"             "LOP(R)"            "SAD(-AMMC)"       
#> [142] "FBl2"              "AME(R)"            "RUB(R)"           
#> [145] "AMMC"              "a2(R)"             "PRW"              
#> [148] "bL(L)"             "PB(R4)"            "a3(R)"            
#> [151] "vACA(R)"          
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                                
#> [1,] "GET"     "/api/v:ver/cached/roicompleteness" 
#> [2,] "GET"     "/api/dbmeta/instances"             
#> [3,] "POST"    "/api/v:ver/npexplorer/roisinneuron"
#> [4,] "REPORT"  "/api/*"                            
#> [5,] "POST"    "/api/v:ver/npexplorer/neuronmeta"  
#> [6,] "OPTIONS" "/api"                              
#>      name                                                                                 
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                 
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDataInstances-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                 
#> [4,] "github.com/labstack/echo.glob..func1"                                               
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                 
#> [6,] "github.com/labstack/echo.glob..func1"                                               

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }