Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "LAL(-GA)(R)"       "AL-DP1m(L)"        "b'L(L)"           
#>   [4] "AL-DM4(R)"         "AL-VM7d(R)"        "NO3(R)"           
#>   [7] "AL-DA3(R)"         "AL-VM7v(R)"        "SMP(L)"           
#>  [10] "BU(L)"             "PVLP(R)"           "OL(R)"            
#>  [13] "AL-DM2(R)"         "EBr3pw"            "AL-VP2(R)"        
#>  [16] "EBr5"              "PB(R8)"            "BU(R)"            
#>  [19] "AL-VC2(R)"         "AL-DA4m(L)"        "AL-DC3(R)"        
#>  [22] "PENP"              "dACA(R)"           "LAL(L)"           
#>  [25] "AOTU(R)"           "PB(L7)"            "AL-VA6(R)"        
#>  [28] "AL-VL1(R)"         "b'2(R)"            "SAD"              
#>  [31] "LH(R)"             "AVLP(R)"           "AB(R)"            
#>  [34] "AL-VA1v(R)"        "AL-VP3(R)"         "PB(L5)"           
#>  [37] "PB(R6)"            "b'L(R)"            "NO(R)"            
#>  [40] "gL(R)"             "CAN(R)"            "SPS(L)"           
#>  [43] "AL-DM1(L)"         "AL-VA7l(R)"        "b2(R)"            
#>  [46] "AL-DP1m(R)"        "EBr6"              "AL-DC2(R)"        
#>  [49] "AMMC"              "FB-column3"        "PB(R4)"           
#>  [52] "AL(R)"             "AL-DP1l(R)"        "g4(R)"            
#>  [55] "PB"                "PLP(R)"            "CRE(L)"           
#>  [58] "LO(R)"             "FBl6"              "AL-DC4(R)"        
#>  [61] "AL-DL1(R)"         "GOR(R)"            "SMP(R)"           
#>  [64] "ME(R)"             "FBl8"              "AL-DL2d(R)"       
#>  [67] "AL-D(R)"           "AL-VA2(R)"         "SLP(R)"           
#>  [70] "CX"                "NO1(L)"            "MB(L)"            
#>  [73] "AL-VL2a(R)"        "ATL(L)"            "g3(R)"            
#>  [76] "EBr2r4"            "AL-DM5(L)"         "AL-VC5(R)"        
#>  [79] "AL(L)"             "a'1(R)"            "a'L(L)"           
#>  [82] "AB(L)"             "a'3(R)"            "g1(R)"            
#>  [85] "ROB(R)"            "PB(R3)"            "ICL(R)"           
#>  [88] "AL-VC1(R)"         "SCL(R)"            "AL-VM7v(L)"       
#>  [91] "AME(R)"            "AL-DA2(R)"         "AL-VC4(R)"        
#>  [94] "AL-D(L)"           "LAL(R)"            "AL-VM5d(R)"       
#>  [97] "CA(L)"             "hemibrain"         "AL-DC2(L)"        
#> [100] "AL-DM3(R)"         "AL-VC3m(R)"        "b'1(R)"           
#> [103] "a2(R)"             "PB(R9)"            "FBl2"             
#> [106] "AL-DA4l(R)"        "AL-VP4(R)"         "SIP(L)"           
#> [109] "FBl1"              "AL-DL4(R)"         "POC"              
#> [112] "a1(R)"             "a3(R)"             "MB(R)"            
#> [115] "AL-VL2p(R)"        "AL-VM4(R)"         "NO2(L)"           
#> [118] "SNP(R)"            "AL-DM6(L)"         "NO3(L)"           
#> [121] "VES(R)"            "RUB(L)"            "FB"               
#> [124] "g2(R)"             "EPA(R)"            "PB(R1)"           
#> [127] "b1(R)"             "VLNP(R)"           "VMNP"             
#> [130] "AL-DA3(L)"         "CRE(-RUB)(L)"      "AL-DL2v(R)"       
#> [133] "AL-DM5(R)"         "GF(R)"             "GOR(L)"           
#> [136] "FLA(R)"            "PB(L4)"            "AL-DA4m(R)"       
#> [139] "AL-DC1(L)"         "ICL(L)"            "NO(L)"            
#> [142] "PB(R5)"            "IPS(R)"            "VES(L)"           
#> [145] "AL-VA7m(R)"        "RUB(R)"            "mALT(R)"          
#> [148] "EBr3d"             "SIP(R)"            "FBl9"             
#> [151] "NO2(R)"            "GC"                "WED(R)"           
#> [154] "AL-DM3(L)"         "PB(L6)"            "FBl4"             
#> [157] "AL-VA3(R)"         "PRW"               "AL-DM4(L)"        
#> [160] "NO"                "FBl3"              "AL-DL5(R)"        
#> [163] "EB"                "AL-DM2(L)"         "aL(L)"            
#> [166] "GNG"               "CRE(R)"            "EBr1"             
#> [169] "gL(L)"             "AL-VA4(R)"         "AL-VA5(R)"        
#> [172] "MB(+ACA)(R)"       "LX(L)"             "AL-DA1(R)"        
#> [175] "AL-DL3(R)"         "AL-VM5v(R)"        "a'L(R)"           
#> [178] "PB(R2)"            "IB"                "g5(R)"            
#> [181] "EPA(L)"            "AL-VP5(R)"         "mALT(L)"          
#> [184] "LOP(R)"            "SPS(R)"            "PB(R7)"           
#> [187] "PED(R)"            "NO1(R)"            "CA(R)"            
#> [190] "PB(L3)"            "a'2(R)"            "IB112"            
#> [193] "AL-V(R)"           "GA(R)"             "ATL(R)"           
#> [196] "AL-VA6(L)"         "AL-VM7d(L)"        "AL-VM1(R)"        
#> [199] "AL-VM2(R)"         "AL-DL4(L)"         "AL-DM1(R)"        
#> [202] "PB(L8)"            "LX(R)"             "INP"              
#> [205] "EBr3am"            "FBl7"              "AL-VP1m(R)"       
#> [208] "SCL(L)"            "CRE(-ROB,-RUB)(R)" "aL(R)"            
#> [211] "AL-VA1d(R)"        "PB(L2)"            "SNP(L)"           
#> [214] "SAD(-AMMC)"        "AL-VC3l(R)"        "AOT(R)"           
#> [217] "AL-DA2(L)"         "FBl5"              "AL-VP1l(R)"       
#> [220] "lACA(R)"           "PB(L1)"            "PB(L9)"           
#> [223] "AL-DM6(R)"         "bL(L)"             "AL-DC1(R)"        
#> [226] "vACA(R)"           "AL-DC4(L)"         "AL-DL5(L)"        
#> [229] "bL(R)"             "AL-VM3(R)"         "AL-VP1d(R)"       
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "a'2(R)"            "PB(L6)"            "gL(R)"            
#>   [4] "FBl2"              "VES(L)"            "GA(R)"            
#>   [7] "hemibrain"         "GC"                "EBr3am"           
#>  [10] "PRW"               "INP"               "CX"               
#>  [13] "g3(R)"             "GOR(L)"            "CA(R)"            
#>  [16] "BU(R)"             "VES(R)"            "SCL(R)"           
#>  [19] "LX(R)"             "LO(R)"             "FBl3"             
#>  [22] "FBl5"              "AOTU(R)"           "lACA(R)"          
#>  [25] "g1(R)"             "FBl8"              "b1(R)"            
#>  [28] "LX(L)"             "AME(R)"            "aL(L)"            
#>  [31] "VLNP(R)"           "AMMC"              "ME(R)"            
#>  [34] "a'1(R)"            "FBl6"              "FBl9"             
#>  [37] "ICL(R)"            "GNG"               "a3(R)"            
#>  [40] "PB"                "EBr2r4"            "FBl1"             
#>  [43] "NO3(R)"            "SAD(-AMMC)"        "g5(R)"            
#>  [46] "IPS(R)"            "bL(R)"             "b'1(R)"           
#>  [49] "b'2(R)"            "SIP(L)"            "EBr5"             
#>  [52] "EBr3d"             "b'L(L)"            "LAL(R)"           
#>  [55] "AOT(R)"            "SMP(R)"            "SMP(L)"           
#>  [58] "NO1(L)"            "PB(L9)"            "ATL(R)"           
#>  [61] "AVLP(R)"           "NO2(R)"            "PB(R6)"           
#>  [64] "EBr6"              "LH(R)"             "POC"              
#>  [67] "mALT(L)"           "NO2(L)"            "b'L(R)"           
#>  [70] "g4(R)"             "NO(L)"             "LOP(R)"           
#>  [73] "PB(L3)"            "a2(R)"             "LAL(-GA)(R)"      
#>  [76] "CRE(-ROB,-RUB)(R)" "PENP"              "EPA(R)"           
#>  [79] "PB(L5)"            "gL(L)"             "MB(R)"            
#>  [82] "RUB(R)"            "PB(R7)"            "BU(L)"            
#>  [85] "EB"                "PB(L7)"            "b2(R)"            
#>  [88] "SNP(R)"            "NO(R)"             "a1(R)"            
#>  [91] "a'L(R)"            "PB(R5)"            "AB(L)"            
#>  [94] "MB(+ACA)(R)"       "PB(L2)"            "ROB(R)"           
#>  [97] "AL(R)"             "ATL(L)"            "PB(L8)"           
#> [100] "NO1(R)"            "PLP(R)"            "PB(R2)"           
#> [103] "PVLP(R)"           "SLP(R)"            "GF(R)"            
#> [106] "mALT(R)"           "PB(L1)"            "PED(R)"           
#> [109] "CRE(L)"            "ICL(L)"            "vACA(R)"          
#> [112] "PB(L4)"            "OL(R)"             "SCL(L)"           
#> [115] "AB(R)"             "LAL(L)"            "aL(R)"            
#> [118] "MB(L)"             "GOR(R)"            "CRE(R)"           
#> [121] "SIP(R)"            "SNP(L)"            "VMNP"             
#> [124] "NO3(L)"            "PB(R3)"            "CAN(R)"           
#> [127] "EPA(L)"            "SPS(R)"            "EBr1"             
#> [130] "dACA(R)"           "RUB(L)"            "PB(R4)"           
#> [133] "SAD"               "FLA(R)"            "AL(L)"            
#> [136] "NO"                "EBr3pw"            "FB"               
#> [139] "g2(R)"             "PB(R8)"            "a'L(L)"           
#> [142] "CA(L)"             "WED(R)"            "PB(R9)"           
#> [145] "PB(R1)"            "bL(L)"             "FBl7"             
#> [148] "IB"                "SPS(L)"            "FBl4"             
#> [151] "a'3(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "VES(R)"            "NO1(L)"            "AL-DL4(L)"        
#>   [4] "GOR(R)"            "AL-VA1v(R)"        "PB(R5)"           
#>   [7] "CX"                "SCL(R)"            "NO1(R)"           
#>  [10] "SAD(-AMMC)"        "AL-DC4(R)"         "EBr3am"           
#>  [13] "NO3(R)"            "AL-VP4(R)"         "AL-DM6(R)"        
#>  [16] "AL-VP5(R)"         "NO(R)"             "a'1(R)"           
#>  [19] "FBl4"              "VES(L)"            "AL(R)"            
#>  [22] "SMP(R)"            "mALT(L)"           "g3(R)"            
#>  [25] "WED(R)"            "gL(L)"             "AB(R)"            
#>  [28] "AL-DL3(R)"         "AL-D(R)"           "AL-VC3m(R)"       
#>  [31] "PB(R6)"            "SNP(L)"            "AL-VL2a(R)"       
#>  [34] "AL-VP1m(R)"        "a2(R)"             "PB(R1)"           
#>  [37] "AL-DL5(R)"         "AL-DM3(R)"         "AL-DP1m(R)"       
#>  [40] "AL-VA6(R)"         "AL-VP3(R)"         "EPA(L)"           
#>  [43] "SLP(R)"            "AL-VC5(R)"         "AL-VM2(R)"        
#>  [46] "LAL(L)"            "g2(R)"             "EBr2r4"           
#>  [49] "AL-DP1m(L)"        "FBl3"              "b'L(L)"           
#>  [52] "MB(R)"             "AL-VA7m(R)"        "SPS(R)"           
#>  [55] "ICL(L)"            "PED(R)"            "EBr1"             
#>  [58] "AL-VA6(L)"         "PB(L6)"            "FBl6"             
#>  [61] "AL-DA3(R)"         "AL-VC4(R)"         "PRW"              
#>  [64] "FBl1"              "AL-DP1l(R)"        "IB"               
#>  [67] "aL(L)"             "bL(R)"             "a'3(R)"           
#>  [70] "EBr3d"             "AME(R)"            "AL-DC1(R)"        
#>  [73] "AL-VL2p(R)"        "AL-VM1(R)"         "PB(R9)"           
#>  [76] "mALT(R)"           "AL-DC3(R)"         "AL-DL2d(R)"       
#>  [79] "NO2(L)"            "PB(R7)"            "AL-DL5(L)"        
#>  [82] "MB(+ACA)(R)"       "NO(L)"             "EB"               
#>  [85] "LO(R)"             "AL-VP1d(R)"        "g1(R)"            
#>  [88] "AL-DC1(L)"         "AL-DM4(L)"         "ROB(R)"           
#>  [91] "CAN(R)"            "AL-VM7v(R)"        "AMMC"             
#>  [94] "GF(R)"             "PB(L8)"            "IPS(R)"           
#>  [97] "AL-DL2v(R)"        "AL-VP1l(R)"        "PB(L2)"           
#> [100] "AL-DM2(R)"         "AL-DA4m(L)"        "AL-DM5(L)"        
#> [103] "LH(R)"             "g4(R)"             "AL-DM2(L)"        
#> [106] "PB(L9)"            "RUB(R)"            "CA(R)"            
#> [109] "AB(L)"             "vACA(R)"           "ME(R)"            
#> [112] "INP"               "SCL(L)"            "LOP(R)"           
#> [115] "GA(R)"             "SPS(L)"            "a'L(R)"           
#> [118] "AL-D(L)"           "PVLP(R)"           "AL-VA5(R)"        
#> [121] "AL-V(R)"           "aL(R)"             "FB"               
#> [124] "hemibrain"         "AL-VM7d(L)"        "LX(L)"            
#> [127] "AL-DL1(R)"         "VLNP(R)"           "SIP(L)"           
#> [130] "AL-DC2(L)"         "FB-column3"        "AL-VC2(R)"        
#> [133] "b2(R)"             "AL-DA2(L)"         "AL-DM6(L)"        
#> [136] "FBl2"              "LAL(R)"            "NO3(L)"           
#> [139] "AL-DC2(R)"         "PB(L1)"            "AL-DM3(L)"        
#> [142] "SIP(R)"            "a'L(L)"            "AL-VA3(R)"        
#> [145] "AL-VA7l(R)"        "PB"                "GNG"              
#> [148] "AL-VC1(R)"         "EBr3pw"            "b1(R)"            
#> [151] "AL-VL1(R)"         "GC"                "EPA(R)"           
#> [154] "bL(L)"             "AL-DA4m(R)"        "AL-VM5d(R)"       
#> [157] "g5(R)"             "FBl9"              "a3(R)"            
#> [160] "MB(L)"             "BU(L)"             "AL-DM4(R)"        
#> [163] "AOT(R)"            "CRE(-ROB,-RUB)(R)" "ATL(L)"           
#> [166] "GOR(L)"            "FLA(R)"            "AL-DM1(L)"        
#> [169] "FBl5"              "NO2(R)"            "AL-VP2(R)"        
#> [172] "LAL(-GA)(R)"       "AL-DA3(L)"         "ATL(R)"           
#> [175] "PB(R3)"            "AL-DA1(R)"         "AL-VA1d(R)"       
#> [178] "CRE(R)"            "PB(R8)"            "AL-VM3(R)"        
#> [181] "AL-VM7d(R)"        "gL(R)"             "lACA(R)"          
#> [184] "FBl8"              "PB(R4)"            "EBr5"             
#> [187] "PB(L7)"            "OL(R)"             "AL-VA2(R)"        
#> [190] "AOTU(R)"           "dACA(R)"           "SNP(R)"           
#> [193] "AL-DA4l(R)"        "AL(L)"             "SMP(L)"           
#> [196] "VMNP"              "LX(R)"             "AL-DC4(L)"        
#> [199] "AL-DA2(R)"         "AL-DM5(R)"         "AL-VM5v(R)"       
#> [202] "POC"               "CA(L)"             "b'L(R)"           
#> [205] "a'2(R)"            "PLP(R)"            "CRE(L)"           
#> [208] "CRE(-RUB)(L)"      "b'2(R)"            "AL-VM7v(L)"       
#> [211] "ICL(R)"            "AL-VM4(R)"         "BU(R)"            
#> [214] "b'1(R)"            "PB(L5)"            "EBr6"             
#> [217] "FBl7"              "PB(R2)"            "NO"               
#> [220] "AL-VC3l(R)"        "a1(R)"             "PENP"             
#> [223] "AL-DL4(R)"         "AL-DM1(R)"         "PB(L3)"           
#> [226] "SAD"               "PB(L4)"            "AL-VA4(R)"        
#> [229] "AVLP(R)"           "RUB(L)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "b'L(L)"            "VES(L)"            "EB"               
#>   [4] "AL-VM7v(L)"        "AL-VL2a(R)"        "EBr3d"            
#>   [7] "AL-VM7d(R)"        "AB(L)"             "SIP(R)"           
#>  [10] "AL-VA3(R)"         "AL-VM5d(R)"        "PRW"              
#>  [13] "PB(R3)"            "LO(R)"             "SLP(R)"           
#>  [16] "AL-D(R)"           "VLNP(R)"           "AL-VA6(L)"        
#>  [19] "AL-DC3(R)"         "AL-DL1(R)"         "AL-VP1l(R)"       
#>  [22] "AL-DA3(L)"         "AL-DM1(R)"         "AL-DM6(R)"        
#>  [25] "AL-VC3l(R)"        "SMP(R)"            "PB(R6)"           
#>  [28] "OL(R)"             "AL-VM4(R)"         "AL-VP4(R)"        
#>  [31] "AL-DC4(R)"         "PVLP(R)"           "AVLP(R)"          
#>  [34] "RUB(L)"            "CRE(-ROB,-RUB)(R)" "SMP(L)"           
#>  [37] "hemibrain"         "FBl2"              "bL(R)"            
#>  [40] "g5(R)"             "aL(L)"             "AL-DM2(R)"        
#>  [43] "PB(L5)"            "gL(L)"             "NO2(R)"           
#>  [46] "CAN(R)"            "AL-DP1l(R)"        "AL-DM3(R)"        
#>  [49] "PENP"              "MB(L)"             "BU(L)"            
#>  [52] "NO3(L)"            "AL-DA4l(R)"        "AL-DA4m(R)"       
#>  [55] "SIP(L)"            "b2(R)"             "AL-VM7d(L)"       
#>  [58] "a'1(R)"            "BU(R)"             "AL-DM3(L)"        
#>  [61] "CRE(L)"            "AL-VA5(R)"         "LAL(-GA)(R)"      
#>  [64] "AL-DM1(L)"         "PB(R1)"            "AL-VC4(R)"        
#>  [67] "AL-DP1m(R)"        "AL-VM3(R)"         "AL-VM5v(R)"       
#>  [70] "NO1(R)"            "PB(L6)"            "AL-DL2v(R)"       
#>  [73] "AL-DL5(R)"         "PB(L8)"            "ME(R)"            
#>  [76] "FBl9"              "CRE(-RUB)(L)"      "ICL(R)"           
#>  [79] "AL-DA2(L)"         "AL-DL2d(R)"        "AL-VL1(R)"        
#>  [82] "CX"                "PB(R5)"            "g2(R)"            
#>  [85] "SPS(L)"            "FBl6"              "LX(L)"            
#>  [88] "AL-D(L)"           "AL-VP1m(R)"        "AL-VP1d(R)"       
#>  [91] "PB(L3)"            "AL-VP3(R)"         "LX(R)"            
#>  [94] "AL-DP1m(L)"        "POC"               "EPA(L)"           
#>  [97] "AL-V(R)"           "SPS(R)"            "GOR(L)"           
#> [100] "WED(R)"            "PB(L9)"            "PB(R8)"           
#> [103] "AL-VA2(R)"         "b'1(R)"            "MB(+ACA)(R)"      
#> [106] "LAL(R)"            "b'L(R)"            "a'2(R)"           
#> [109] "g4(R)"             "AL-VC1(R)"         "SAD"              
#> [112] "b1(R)"             "CA(L)"             "EBr6"             
#> [115] "IPS(R)"            "AOT(R)"            "AMMC"             
#> [118] "PB(R9)"            "AL-DM5(R)"         "AL-VA1v(R)"       
#> [121] "NO2(L)"            "GNG"               "AL(R)"            
#> [124] "AL-DC1(R)"         "AL-DM4(R)"         "NO3(R)"           
#> [127] "EBr2r4"            "gL(R)"             "CA(R)"            
#> [130] "PB(L2)"            "EPA(R)"            "AL-VC5(R)"        
#> [133] "SCL(L)"            "LOP(R)"            "aL(R)"            
#> [136] "SCL(R)"            "SNP(R)"            "AME(R)"           
#> [139] "NO"                "PB(R4)"            "AL-VM1(R)"        
#> [142] "a'L(R)"            "a2(R)"             "mALT(R)"          
#> [145] "AL-DM6(L)"         "INP"               "AL-DL4(R)"        
#> [148] "g3(R)"             "g1(R)"             "SNP(L)"           
#> [151] "LH(R)"             "FBl8"              "AL-VA7m(R)"       
#> [154] "RUB(R)"            "AL-DL5(L)"         "FBl1"             
#> [157] "AL-DA1(R)"         "AL-VC2(R)"         "AL-DM4(L)"        
#> [160] "a3(R)"             "EBr3am"            "FBl3"             
#> [163] "a'L(L)"            "FBl4"              "ATL(R)"           
#> [166] "GOR(R)"            "vACA(R)"           "FB"               
#> [169] "dACA(R)"           "FLA(R)"            "AL-DM5(L)"        
#> [172] "LAL(L)"            "AL-VM2(R)"         "AL-DA3(R)"        
#> [175] "AOTU(R)"           "AL(L)"             "NO(R)"            
#> [178] "FBl5"              "AL-VC3m(R)"        "NO1(L)"           
#> [181] "EBr3pw"            "VES(R)"            "AL-VA7l(R)"       
#> [184] "PB(L1)"            "PB(L7)"            "ATL(L)"           
#> [187] "EBr5"              "PB(L4)"            "AL-DA4m(L)"       
#> [190] "AB(R)"             "b'2(R)"            "CRE(R)"           
#> [193] "a'3(R)"            "a1(R)"             "AL-VP5(R)"        
#> [196] "AL-VP2(R)"         "ROB(R)"            "FB-column3"       
#> [199] "NO(L)"             "AL-VL2p(R)"        "AL-VA6(R)"        
#> [202] "PED(R)"            "EBr1"              "ICL(L)"           
#> [205] "AL-VA4(R)"         "GF(R)"             "FBl7"             
#> [208] "AL-VM7v(R)"        "lACA(R)"           "VMNP"             
#> [211] "bL(L)"             "SAD(-AMMC)"        "AL-DC4(L)"        
#> [214] "AL-DL4(L)"         "AL-VA1d(R)"        "IB"               
#> [217] "GA(R)"             "PB(R7)"            "PB"               
#> [220] "AL-DC1(L)"         "mALT(L)"           "AL-DC2(L)"        
#> [223] "AL-DC2(R)"         "MB(R)"             "PLP(R)"           
#> [226] "AL-DM2(L)"         "GC"                "PB(R2)"           
#> [229] "AL-DA2(R)"         "AL-DL3(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method  path                                
#> [1,] "PUT"   "/api/*"                            
#> [2,] "TRACE" "/api/*"                            
#> [3,] "GET"   "/api/v:ver/cached/dailytype"       
#> [4,] "POST"  "/api/npexplorer/roisinneuron"      
#> [5,] "POST"  "/api/npexplorer/commonconnectivity"
#> [6,] "POST"  "/api/npexplorer/autapses"          
#>      name                                                                                              
#> [1,] "github.com/labstack/echo.glob..func1"                                                            
#> [2,] "github.com/labstack/echo.glob..func1"                                                            
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"      
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCommonConnectivity-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"          

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }