Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DL2v(R)"        "AB(R)"             "NO2(R)"           
#>   [4] "AL-VM3(R)"         "b'1(R)"            "GA(R)"            
#>   [7] "FB"                "g2(R)"             "SNP(L)"           
#>  [10] "IB112"             "AL-DA2(R)"         "PB(L7)"           
#>  [13] "bL(R)"             "AL-VP5(R)"         "AL-DC3(R)"        
#>  [16] "SPS(R)"            "MB(L)"             "AL-DC4(R)"        
#>  [19] "PLP(R)"            "gL(R)"             "ICL(L)"           
#>  [22] "AL-VC4(R)"         "AL-VM4(R)"         "EB"               
#>  [25] "AL-D(L)"           "LH(R)"             "VES(L)"           
#>  [28] "GOR(L)"            "AL-DL4(L)"         "AL-VP1m(R)"       
#>  [31] "AL(L)"             "AL-DC4(L)"         "AL(R)"            
#>  [34] "AL-VM5v(R)"        "a1(R)"             "b2(R)"            
#>  [37] "FBl3"              "FBl7"              "AL-DP1m(R)"       
#>  [40] "AL-VM1(R)"         "a'3(R)"            "SAD"              
#>  [43] "PB(L4)"            "AL-VM2(R)"         "LOP(R)"           
#>  [46] "CRE(R)"            "AL-DM6(L)"         "NO"               
#>  [49] "AL-VL2a(R)"        "AL-VP1d(R)"        "IB"               
#>  [52] "MB(R)"             "LX(R)"             "LAL(R)"           
#>  [55] "AL-VC5(R)"         "ATL(L)"            "PB(L9)"           
#>  [58] "AL-DL4(R)"         "LAL(L)"            "EBr5"             
#>  [61] "SIP(R)"            "AL-DM3(L)"         "AL-VA5(R)"        
#>  [64] "EBr3d"             "a'1(R)"            "CA(R)"            
#>  [67] "PB(R2)"            "AL-VL1(R)"         "CX"               
#>  [70] "PED(R)"            "AL-VC1(R)"         "AL-VP3(R)"        
#>  [73] "AL-DM2(R)"         "AL-VM7d(R)"        "SCL(R)"           
#>  [76] "AL-DM2(L)"         "AME(R)"            "GNG"              
#>  [79] "GOR(R)"            "a'L(L)"            "VMNP"             
#>  [82] "AL-DM1(L)"         "AVLP(R)"           "OL(R)"            
#>  [85] "PB(L2)"            "INP"               "LAL(-GA)(R)"      
#>  [88] "dACA(R)"           "CRE(L)"            "SLP(R)"           
#>  [91] "AL-DP1l(R)"        "AL-VA2(R)"         "g1(R)"            
#>  [94] "AL-VM7v(L)"        "GF(R)"             "SMP(R)"           
#>  [97] "RUB(R)"            "b'L(R)"            "g5(R)"            
#> [100] "FBl6"              "b1(R)"             "AL-V(R)"          
#> [103] "VLNP(R)"           "NO3(R)"            "PB(L1)"           
#> [106] "SNP(R)"            "PB(R9)"            "AL-D(R)"          
#> [109] "AL-VP2(R)"         "vACA(R)"           "SMP(L)"           
#> [112] "NO1(R)"            "PB"                "VES(R)"           
#> [115] "PB(R5)"            "AB(L)"             "hemibrain"        
#> [118] "NO1(L)"            "a'2(R)"            "FB-column3"       
#> [121] "AL-VP1l(R)"        "g3(R)"             "PB(L8)"           
#> [124] "MB(+ACA)(R)"       "AL-VP4(R)"         "AL-VL2p(R)"       
#> [127] "SIP(L)"            "EPA(R)"            "LX(L)"            
#> [130] "EPA(L)"            "AL-DL3(R)"         "AL-DL5(R)"        
#> [133] "AL-VA6(R)"         "lACA(R)"           "WED(R)"           
#> [136] "AL-DA4m(L)"        "CAN(R)"            "RUB(L)"           
#> [139] "AL-DA3(R)"         "GC"                "EBr1"             
#> [142] "CA(L)"             "mALT(R)"           "FBl8"             
#> [145] "AL-DM1(R)"         "g4(R)"             "AL-VC3m(R)"       
#> [148] "PB(R8)"            "NO(R)"             "AL-DC1(R)"        
#> [151] "EBr3am"            "PB(R1)"            "AL-DA3(L)"        
#> [154] "NO3(L)"            "FBl4"              "AL-DM4(L)"        
#> [157] "ICL(R)"            "CRE(-RUB)(L)"      "AL-DA4l(R)"       
#> [160] "AL-VA1d(R)"        "b'2(R)"            "PRW"              
#> [163] "FLA(R)"            "AL-DA1(R)"         "AOT(R)"           
#> [166] "PB(L5)"            "a2(R)"             "PB(R4)"           
#> [169] "SCL(L)"            "EBr2r4"            "PB(R3)"           
#> [172] "AL-VC2(R)"         "PB(R6)"            "AL-VA6(L)"        
#> [175] "bL(L)"             "POC"               "AL-DC2(L)"        
#> [178] "a3(R)"             "AL-DM3(R)"         "aL(R)"            
#> [181] "gL(L)"             "AL-VA1v(R)"        "AL-DA2(L)"        
#> [184] "FBl9"              "AL-DM4(R)"         "AL-VM7d(L)"       
#> [187] "IPS(R)"            "AL-DA4m(R)"        "AL-VA3(R)"        
#> [190] "ME(R)"             "BU(R)"             "NO2(L)"           
#> [193] "PB(R7)"            "EBr6"              "BU(L)"            
#> [196] "AL-DL5(L)"         "NO(L)"             "AL-VM7v(R)"       
#> [199] "PENP"              "AOTU(R)"           "LO(R)"            
#> [202] "AL-DC2(R)"         "AL-DM6(R)"         "AL-VA7m(R)"       
#> [205] "AL-DP1m(L)"        "AL-DL1(R)"         "ATL(R)"           
#> [208] "AL-DL2d(R)"        "CRE(-ROB,-RUB)(R)" "mALT(L)"          
#> [211] "AMMC"              "AL-VA4(R)"         "AL-VM5d(R)"       
#> [214] "AL-VA7l(R)"        "a'L(R)"            "ROB(R)"           
#> [217] "PB(L6)"            "PVLP(R)"           "FBl5"             
#> [220] "b'L(L)"            "PB(L3)"            "AL-DM5(L)"        
#> [223] "FBl1"              "FBl2"              "AL-DM5(R)"        
#> [226] "AL-VC3l(R)"        "SPS(L)"            "AL-DC1(L)"        
#> [229] "EBr3pw"            "SAD(-AMMC)"        "aL(L)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "WED(R)"            "CAN(R)"            "GNG"              
#>   [4] "VLNP(R)"           "mALT(L)"           "ATL(L)"           
#>   [7] "CX"                "PB(R7)"            "b'L(R)"           
#>  [10] "EBr2r4"            "AB(R)"             "SMP(R)"           
#>  [13] "IPS(R)"            "OL(R)"             "LOP(R)"           
#>  [16] "EBr1"              "gL(L)"             "NO"               
#>  [19] "PVLP(R)"           "ATL(R)"            "SMP(L)"           
#>  [22] "PB(L3)"            "LH(R)"             "SPS(R)"           
#>  [25] "FBl6"              "AVLP(R)"           "b1(R)"            
#>  [28] "RUB(L)"            "a'L(L)"            "CA(L)"            
#>  [31] "EBr5"              "INP"               "MB(L)"            
#>  [34] "FLA(R)"            "g5(R)"             "LAL(R)"           
#>  [37] "EBr3am"            "SCL(L)"            "PB(L2)"           
#>  [40] "g1(R)"             "aL(R)"             "FB"               
#>  [43] "BU(L)"             "IB"                "ICL(R)"           
#>  [46] "b'2(R)"            "NO3(R)"            "PB(R1)"           
#>  [49] "ICL(L)"            "SNP(L)"            "PB(R6)"           
#>  [52] "PB(R8)"            "EPA(R)"            "PB(R9)"           
#>  [55] "ROB(R)"            "NO1(L)"            "FBl7"             
#>  [58] "EBr6"              "LX(R)"             "BU(R)"            
#>  [61] "MB(R)"             "PB(L5)"            "AOTU(R)"          
#>  [64] "GC"                "PLP(R)"            "ME(R)"            
#>  [67] "PB(L4)"            "FBl2"              "CA(R)"            
#>  [70] "vACA(R)"           "GF(R)"             "SAD"              
#>  [73] "PB(L1)"            "SNP(R)"            "SIP(R)"           
#>  [76] "AMMC"              "a2(R)"             "PB(L6)"           
#>  [79] "gL(R)"             "EBr3pw"            "PB(R4)"           
#>  [82] "mALT(R)"           "CRE(R)"            "SCL(R)"           
#>  [85] "NO1(R)"            "LX(L)"             "VES(L)"           
#>  [88] "RUB(R)"            "b2(R)"             "EB"               
#>  [91] "CRE(L)"            "NO2(R)"            "NO(L)"            
#>  [94] "FBl3"              "b'L(L)"            "a'3(R)"           
#>  [97] "lACA(R)"           "g2(R)"             "dACA(R)"          
#> [100] "VMNP"              "AL(L)"             "CRE(-ROB,-RUB)(R)"
#> [103] "LAL(L)"            "LAL(-GA)(R)"       "a'2(R)"           
#> [106] "NO3(L)"            "PB(L7)"            "PB(L9)"           
#> [109] "LO(R)"             "PB(R3)"            "AME(R)"           
#> [112] "GA(R)"             "bL(L)"             "MB(+ACA)(R)"      
#> [115] "EPA(L)"            "a3(R)"             "FBl8"             
#> [118] "AB(L)"             "a1(R)"             "g3(R)"            
#> [121] "GOR(L)"            "a'L(R)"            "SAD(-AMMC)"       
#> [124] "FBl5"              "PENP"              "FBl9"             
#> [127] "SLP(R)"            "g4(R)"             "aL(L)"            
#> [130] "AL(R)"             "b'1(R)"            "PB(R5)"           
#> [133] "NO2(L)"            "PRW"               "SIP(L)"           
#> [136] "PB(L8)"            "PB"                "FBl4"             
#> [139] "GOR(R)"            "POC"               "hemibrain"        
#> [142] "NO(R)"             "FBl1"              "EBr3d"            
#> [145] "a'1(R)"            "AOT(R)"            "PED(R)"           
#> [148] "SPS(L)"            "VES(R)"            "bL(R)"            
#> [151] "PB(R2)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "hemibrain"         "CRE(R)"            "AL-DA2(L)"        
#>   [4] "AL-DA2(R)"         "AL-DM6(R)"         "AL-VP3(R)"        
#>   [7] "AL-VP1m(R)"        "EBr3pw"            "a1(R)"            
#>  [10] "PRW"               "GOR(L)"            "mALT(R)"          
#>  [13] "aL(R)"             "EPA(R)"            "AL-DL5(L)"        
#>  [16] "LO(R)"             "AL-DP1l(R)"        "EBr2r4"           
#>  [19] "SAD(-AMMC)"        "AL-DA1(R)"         "AL-VA4(R)"        
#>  [22] "AL-VM3(R)"         "LAL(-GA)(R)"       "FBl7"             
#>  [25] "BU(R)"             "g3(R)"             "g1(R)"            
#>  [28] "PB(R8)"            "AL-VA7m(R)"        "CRE(-ROB,-RUB)(R)"
#>  [31] "AOTU(R)"           "FBl3"              "VES(L)"           
#>  [34] "AL-VP1d(R)"        "PB(R3)"            "GNG"              
#>  [37] "a'L(R)"            "AL-DC1(L)"         "AL-DC2(R)"        
#>  [40] "AL-DL3(R)"         "a'3(R)"            "mALT(L)"          
#>  [43] "AL-DA3(L)"         "NO(R)"             "PB(L9)"           
#>  [46] "FBl4"              "AL-DM4(R)"         "AL-VM4(R)"        
#>  [49] "SMP(R)"            "FLA(R)"            "AL-DM2(L)"        
#>  [52] "MB(+ACA)(R)"       "NO2(R)"            "AL-VA3(R)"        
#>  [55] "AL-VM7d(R)"        "AL-VP4(R)"         "PB(L3)"           
#>  [58] "CRE(L)"            "AL-VC2(R)"         "AL-VC4(R)"        
#>  [61] "LX(R)"             "PVLP(R)"           "AL-DC1(R)"        
#>  [64] "g2(R)"             "AB(L)"             "lACA(R)"          
#>  [67] "EB"                "SIP(R)"            "PB(L6)"           
#>  [70] "gL(R)"             "FB-column3"        "PB(R4)"           
#>  [73] "PED(R)"            "dACA(R)"           "AL-VA1d(R)"       
#>  [76] "AL-VC3m(R)"        "BU(L)"             "NO1(R)"           
#>  [79] "AL-DL1(R)"         "SMP(L)"            "AL-DA3(R)"        
#>  [82] "AL-DM3(R)"         "FBl9"              "ICL(R)"           
#>  [85] "LX(L)"             "AL-VL2a(R)"        "AL-VM5d(R)"       
#>  [88] "EBr3am"            "g5(R)"             "a2(R)"            
#>  [91] "SPS(R)"            "AL-DM5(R)"         "AL-DM4(L)"        
#>  [94] "CA(R)"             "RUB(R)"            "SLP(R)"           
#>  [97] "AL-DL4(L)"         "FBl6"              "SIP(L)"           
#> [100] "a3(R)"             "FBl2"              "PB(R2)"           
#> [103] "b'L(L)"            "NO(L)"             "a'1(R)"           
#> [106] "aL(L)"             "AOT(R)"            "b'2(R)"           
#> [109] "IB"                "VMNP"              "PB(L7)"           
#> [112] "MB(R)"             "AL-V(R)"           "PB(L1)"           
#> [115] "PB"                "SNP(R)"            "gL(L)"            
#> [118] "CRE(-RUB)(L)"      "GOR(R)"            "AL-VM1(R)"        
#> [121] "AL(L)"             "NO2(L)"            "PB(L8)"           
#> [124] "AL-VM7v(L)"        "AME(R)"            "AL-VC5(R)"        
#> [127] "AL-VP2(R)"         "PENP"              "PB(R1)"           
#> [130] "AL-DL2v(R)"        "EBr3d"             "g4(R)"            
#> [133] "AL-DC4(L)"         "AL-DM1(R)"         "AL-VL1(R)"        
#> [136] "AL-VM2(R)"         "AL-D(L)"           "VLNP(R)"          
#> [139] "GF(R)"             "NO1(L)"            "CAN(R)"           
#> [142] "AL-VA7l(R)"        "PB(R5)"            "PB(L2)"           
#> [145] "FB"                "FBl1"              "EBr6"             
#> [148] "FBl8"              "EPA(L)"            "GC"               
#> [151] "AL(R)"             "PB(L5)"            "EBr5"             
#> [154] "AL-DM1(L)"         "AL-DP1m(L)"        "AL-DL5(R)"        
#> [157] "AL-DL2d(R)"        "AL-VA2(R)"         "LAL(R)"           
#> [160] "LAL(L)"            "AL-DM5(L)"         "AL-DM6(L)"        
#> [163] "LH(R)"             "AL-D(R)"           "ATL(L)"           
#> [166] "AMMC"              "WED(R)"            "b'L(R)"           
#> [169] "a'2(R)"            "AL-DC2(L)"         "SNP(L)"           
#> [172] "AB(R)"             "AL-VP5(R)"         "MB(L)"            
#> [175] "EBr1"              "AL-DA4m(L)"        "NO"               
#> [178] "CA(L)"             "AL-DM3(L)"         "PB(R9)"           
#> [181] "bL(L)"             "bL(R)"             "a'L(L)"           
#> [184] "AL-DC4(R)"         "AL-VC3l(R)"        "AL-VM7d(L)"       
#> [187] "GA(R)"             "ICL(L)"            "IPS(R)"           
#> [190] "POC"               "b'1(R)"            "PB(R7)"           
#> [193] "LOP(R)"            "ME(R)"             "AL-VM7v(R)"       
#> [196] "AL-VP1l(R)"        "SCL(R)"            "INP"              
#> [199] "AL-VA1v(R)"        "RUB(L)"            "OL(R)"            
#> [202] "AL-VA6(L)"         "AL-DP1m(R)"        "AL-VL2p(R)"       
#> [205] "SCL(L)"            "PB(R6)"            "SAD"              
#> [208] "VES(R)"            "b1(R)"             "vACA(R)"          
#> [211] "PLP(R)"            "AL-VM5v(R)"        "SPS(L)"           
#> [214] "NO3(L)"            "AVLP(R)"           "CX"               
#> [217] "NO3(R)"            "FBl5"              "ATL(R)"           
#> [220] "AL-DA4l(R)"        "AL-DC3(R)"         "AL-DM2(R)"        
#> [223] "AL-VC1(R)"         "b2(R)"             "ROB(R)"           
#> [226] "PB(L4)"            "AL-DA4m(R)"        "AL-DL4(R)"        
#> [229] "AL-VA5(R)"         "AL-VA6(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-D(L)"           "AL-DM6(L)"         "NO2(R)"           
#>   [4] "MB(R)"             "AL-DL2v(R)"        "IB"               
#>   [7] "GF(R)"             "PVLP(R)"           "CAN(R)"           
#>  [10] "AL-DA2(L)"         "FBl9"              "a'2(R)"           
#>  [13] "EBr6"              "PB(R8)"            "GOR(L)"           
#>  [16] "EBr1"              "AMMC"              "NO(L)"            
#>  [19] "aL(R)"             "AL-VM7d(L)"        "AL-DP1l(R)"       
#>  [22] "AME(R)"            "AL-DC4(R)"         "PB(L9)"           
#>  [25] "MB(L)"             "AL(R)"             "NO2(L)"           
#>  [28] "GC"                "LOP(R)"            "LX(R)"            
#>  [31] "ICL(L)"            "AL-DL3(R)"         "AL-VA7m(R)"       
#>  [34] "b2(R)"             "dACA(R)"           "RUB(R)"           
#>  [37] "AL(L)"             "FBl3"              "AL-VM7v(L)"       
#>  [40] "PB(R9)"            "CRE(L)"            "AL-DL5(R)"        
#>  [43] "AL-VM4(R)"         "AL-VP1l(R)"        "EBr3d"            
#>  [46] "AL-DP1m(L)"        "ME(R)"             "AB(R)"            
#>  [49] "g3(R)"             "FLA(R)"            "AOT(R)"           
#>  [52] "AL-DA1(R)"         "AL-DC2(R)"         "CRE(-RUB)(L)"     
#>  [55] "INP"               "AL-VA5(R)"         "AB(L)"            
#>  [58] "SMP(R)"            "LAL(-GA)(R)"       "g2(R)"            
#>  [61] "FBl5"              "PB(L8)"            "AL-DM3(L)"        
#>  [64] "POC"               "SPS(R)"            "LAL(R)"           
#>  [67] "ATL(R)"            "AL-DC3(R)"         "AL-VA6(L)"        
#>  [70] "FBl4"              "b'L(L)"            "AL-VA1v(R)"       
#>  [73] "AL-VA2(R)"         "PED(R)"            "AL-DM5(L)"        
#>  [76] "AL-DA2(R)"         "AL-DC1(R)"         "AL-VA1d(R)"       
#>  [79] "a'1(R)"            "CA(R)"             "AL-VC3m(R)"       
#>  [82] "AL-VP5(R)"         "EBr3am"            "a'3(R)"           
#>  [85] "NO3(L)"            "PB(L7)"            "AVLP(R)"          
#>  [88] "PB(R6)"            "AL-DC2(L)"         "gL(R)"            
#>  [91] "g5(R)"             "ATL(L)"            "NO3(R)"           
#>  [94] "NO(R)"             "bL(R)"             "AL-DM4(L)"        
#>  [97] "PB(R4)"            "EBr5"              "VES(R)"           
#> [100] "AL-VC5(R)"         "PLP(R)"            "AL-DL5(L)"        
#> [103] "AL-VP2(R)"         "FB"                "AL-VC2(R)"        
#> [106] "CRE(-ROB,-RUB)(R)" "GA(R)"             "AL-VP1m(R)"       
#> [109] "mALT(R)"           "AL-D(R)"           "AL-VM1(R)"        
#> [112] "PENP"              "AOTU(R)"           "SNP(L)"           
#> [115] "b1(R)"             "SLP(R)"            "CA(L)"            
#> [118] "NO1(R)"            "PB(L6)"            "AL-VA6(R)"        
#> [121] "mALT(L)"           "SAD"               "EB"               
#> [124] "LO(R)"             "AL-DA3(R)"         "AL-DP1m(R)"       
#> [127] "AL-VM7v(R)"        "g1(R)"             "AL-DA4m(L)"       
#> [130] "FBl6"              "LX(L)"             "AL-DL2d(R)"       
#> [133] "AL-DL4(R)"         "WED(R)"            "a'L(R)"           
#> [136] "AL-V(R)"           "AL-DA4m(R)"        "AL-VA4(R)"        
#> [139] "PB(L1)"            "AL-VM5v(R)"        "vACA(R)"          
#> [142] "NO1(L)"            "AL-VL2p(R)"        "b'1(R)"           
#> [145] "AL-VP4(R)"         "FBl2"              "VES(L)"           
#> [148] "AL-DL4(L)"         "FBl1"              "OL(R)"            
#> [151] "CX"                "lACA(R)"           "EBr2r4"           
#> [154] "AL-DM5(R)"         "NO"                "AL-DA4l(R)"       
#> [157] "GOR(R)"            "PB(R5)"            "SNP(R)"           
#> [160] "BU(R)"             "VMNP"              "aL(L)"            
#> [163] "a3(R)"             "AL-VA7l(R)"        "a1(R)"            
#> [166] "CRE(R)"            "b'L(R)"            "ROB(R)"           
#> [169] "SAD(-AMMC)"        "PB(R7)"            "AL-DA3(L)"        
#> [172] "RUB(L)"            "GNG"               "FB-column3"       
#> [175] "IPS(R)"            "BU(L)"             "PB(R1)"           
#> [178] "AL-DM6(R)"         "PB(L3)"            "hemibrain"        
#> [181] "SCL(R)"            "SIP(R)"            "AL-VM2(R)"        
#> [184] "PB(R2)"            "AL-VA3(R)"         "MB(+ACA)(R)"      
#> [187] "AL-VL2a(R)"        "a2(R)"             "AL-DM2(L)"        
#> [190] "AL-VM7d(R)"        "AL-VP3(R)"         "SIP(L)"           
#> [193] "FBl7"              "bL(L)"             "LH(R)"            
#> [196] "AL-DM1(R)"         "AL-DM3(R)"         "AL-DM4(R)"        
#> [199] "AL-VC3l(R)"        "b'2(R)"            "AL-DC4(L)"        
#> [202] "AL-VM5d(R)"        "SPS(L)"            "VLNP(R)"          
#> [205] "AL-DL1(R)"         "AL-VP1d(R)"        "SCL(L)"           
#> [208] "PB"                "a'L(L)"            "AL-VM3(R)"        
#> [211] "g4(R)"             "AL-VL1(R)"         "PB(L5)"           
#> [214] "EPA(R)"            "AL-VC1(R)"         "AL-DC1(L)"        
#> [217] "AL-DM2(R)"         "gL(L)"             "EBr3pw"           
#> [220] "AL-VC4(R)"         "LAL(L)"            "PRW"              
#> [223] "AL-DM1(L)"         "PB(R3)"            "EPA(L)"           
#> [226] "PB(L4)"            "PB(L2)"            "SMP(L)"           
#> [229] "FBl8"              "ICL(R)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                                     
#> [1,] "OPTIONS" "/api"                                   
#> [2,] "POST"    "/api/raw/keyvalue/key/:instance/:key"   
#> [3,] "GET"     "/api/v:ver/npexplorer/distribution"     
#> [4,] "POST"    "/api/v:ver/roimeshes/mesh/:dataset/:roi"
#> [5,] "GET"     "/api/cached/roicompleteness"            
#> [6,] "GET"     "/api/custom/custom"                     
#>      name                                                                                    
#> [1,] "github.com/labstack/echo.glob..func1"                                                  
#> [2,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROICompleteness-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"         

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }