Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "PB(L1)"            "AL-DM5(L)"         "MB(+ACA)(R)"      
#>   [4] "AL-DL2d(R)"        "CRE(-ROB,-RUB)(R)" "NO2(L)"           
#>   [7] "AL-DM4(L)"         "AL-DM2(R)"         "AL-VP2(R)"        
#>  [10] "AVLP(R)"           "AL-VA3(R)"         "FLA(R)"           
#>  [13] "AL-V(R)"           "a2(R)"             "MB(R)"            
#>  [16] "GA(R)"             "EPA(R)"            "AL-DL5(L)"        
#>  [19] "AL-DA2(R)"         "AL-VM4(R)"         "AL-VM7v(R)"       
#>  [22] "NO1(L)"            "AL-VP1d(R)"        "AL-DM2(L)"        
#>  [25] "PB(L7)"            "PB(L5)"            "EBr6"             
#>  [28] "NO(L)"             "AL-VC4(R)"         "g3(R)"            
#>  [31] "GC"                "AL-VM7v(L)"        "AL-DA4m(R)"       
#>  [34] "SCL(R)"            "AL-VA6(L)"         "PB(R1)"           
#>  [37] "RUB(L)"            "AL-DP1m(R)"        "AL-VM2(R)"        
#>  [40] "AL-VM3(R)"         "AB(R)"             "INP"              
#>  [43] "a1(R)"             "SIP(R)"            "ROB(R)"           
#>  [46] "CRE(L)"            "FBl5"              "AL-VA6(R)"        
#>  [49] "ME(R)"             "NO"                "FBl8"             
#>  [52] "AL-DM5(R)"         "AL-DM6(R)"         "FBl4"             
#>  [55] "FBl9"              "b'L(L)"            "RUB(R)"           
#>  [58] "LOP(R)"            "AL-DA4m(L)"        "PB(L9)"           
#>  [61] "b1(R)"             "AME(R)"            "AL-VP4(R)"        
#>  [64] "AMMC"              "g1(R)"             "EBr1"             
#>  [67] "EBr3pw"            "LX(L)"             "mALT(L)"          
#>  [70] "AL-VM7d(L)"        "a3(R)"             "SMP(R)"           
#>  [73] "IB"                "PENP"              "EPA(L)"           
#>  [76] "GNG"               "AL-D(R)"           "AL-VA2(R)"        
#>  [79] "PB(R6)"            "BU(L)"             "PB"               
#>  [82] "LX(R)"             "PB(L3)"            "BU(R)"            
#>  [85] "aL(L)"             "dACA(R)"           "a'2(R)"           
#>  [88] "gL(R)"             "CRE(R)"            "PB(R9)"           
#>  [91] "FBl1"              "AL-DL3(R)"         "AL-VM1(R)"        
#>  [94] "AL-VM7d(R)"        "SMP(L)"            "SIP(L)"           
#>  [97] "PB(L6)"            "g5(R)"             "CAN(R)"           
#> [100] "AB(L)"             "AOTU(R)"           "AL-DM6(L)"        
#> [103] "AOT(R)"            "OL(R)"             "AL-VL2a(R)"       
#> [106] "AL-DC2(L)"         "ICL(L)"            "b'1(R)"           
#> [109] "PB(L2)"            "lACA(R)"           "WED(R)"           
#> [112] "AL-D(L)"           "LAL(L)"            "AL-VA7m(R)"       
#> [115] "a'3(R)"            "GF(R)"             "AL-DC1(L)"        
#> [118] "a'L(L)"            "AL-DC3(R)"         "a'1(R)"           
#> [121] "FBl3"              "AL-DL4(R)"         "AL-DM4(R)"        
#> [124] "LAL(-GA)(R)"       "SAD(-AMMC)"        "GOR(R)"           
#> [127] "SLP(R)"            "a'L(R)"            "ATL(R)"           
#> [130] "AL-DL2v(R)"        "AL-VA4(R)"         "vACA(R)"          
#> [133] "PB(R2)"            "AL-VP1l(R)"        "PB(R8)"           
#> [136] "VES(R)"            "NO2(R)"            "AL-DM3(R)"        
#> [139] "AL-VC2(R)"         "PED(R)"            "b2(R)"            
#> [142] "PLP(R)"            "SNP(R)"            "EB"               
#> [145] "AL(L)"             "AL-DM1(L)"         "bL(L)"            
#> [148] "AL-VL1(R)"         "AL-VP5(R)"         "POC"              
#> [151] "PB(L8)"            "NO3(L)"            "SNP(L)"           
#> [154] "IPS(R)"            "PB(R4)"            "AL-DL5(R)"        
#> [157] "FB-column3"        "AL-DA1(R)"         "AL-DC1(R)"        
#> [160] "EBr3am"            "aL(R)"             "EBr2r4"           
#> [163] "AL-DL1(R)"         "PVLP(R)"           "AL(R)"            
#> [166] "AL-VM5v(R)"        "ATL(L)"            "EBr3d"            
#> [169] "g4(R)"             "SAD"               "AL-DP1l(R)"       
#> [172] "PB(L4)"            "AL-VA1v(R)"        "AL-VP3(R)"        
#> [175] "PB(R5)"            "NO3(R)"            "SPS(L)"           
#> [178] "FBl2"              "AL-VA5(R)"         "VLNP(R)"          
#> [181] "IB112"             "AL-VM5d(R)"        "AL-VC1(R)"        
#> [184] "PB(R7)"            "NO(R)"             "CX"               
#> [187] "FB"                "b'L(R)"            "EBr5"             
#> [190] "AL-DP1m(L)"        "CA(R)"             "ICL(R)"           
#> [193] "AL-DC2(R)"         "AL-DM1(R)"         "SCL(L)"           
#> [196] "VMNP"              "AL-DC4(L)"         "AL-DL4(L)"        
#> [199] "LAL(R)"            "MB(L)"             "FBl6"             
#> [202] "FBl7"              "LH(R)"             "AL-VA7l(R)"       
#> [205] "CA(L)"             "PRW"               "AL-DC4(R)"        
#> [208] "AL-VC3l(R)"        "AL-VC3m(R)"        "AL-VC5(R)"        
#> [211] "mALT(R)"           "AL-DA2(L)"         "LO(R)"            
#> [214] "bL(R)"             "GOR(L)"            "AL-VA1d(R)"       
#> [217] "AL-VP1m(R)"        "b'2(R)"            "AL-DM3(L)"        
#> [220] "g2(R)"             "AL-DA3(L)"         "PB(R3)"           
#> [223] "NO1(R)"            "gL(L)"             "VES(L)"           
#> [226] "AL-DA3(R)"         "AL-DA4l(R)"        "hemibrain"        
#> [229] "SPS(R)"            "CRE(-RUB)(L)"      "AL-VL2p(R)"       
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "VLNP(R)"           "PB(L5)"            "SIP(L)"           
#>   [4] "bL(L)"             "FBl9"              "LH(R)"            
#>   [7] "IB"                "PB(L8)"            "a'1(R)"           
#>  [10] "ICL(L)"            "AL(L)"             "SAD(-AMMC)"       
#>  [13] "FBl4"              "MB(L)"             "bL(R)"            
#>  [16] "vACA(R)"           "SPS(R)"            "g2(R)"            
#>  [19] "NO3(L)"            "SIP(R)"            "AME(R)"           
#>  [22] "POC"               "PENP"              "PLP(R)"           
#>  [25] "SNP(R)"            "hemibrain"         "dACA(R)"          
#>  [28] "FLA(R)"            "SCL(R)"            "PB"               
#>  [31] "a'L(R)"            "AB(L)"             "AOTU(R)"          
#>  [34] "PB(R1)"            "AB(R)"             "PB(R3)"           
#>  [37] "GOR(R)"            "a3(R)"             "EBr3am"           
#>  [40] "a1(R)"             "GC"                "BU(R)"            
#>  [43] "NO"                "NO2(L)"            "FBl7"             
#>  [46] "PB(L4)"            "CA(L)"             "SAD"              
#>  [49] "SPS(L)"            "EPA(R)"            "AL(R)"            
#>  [52] "GF(R)"             "RUB(L)"            "MB(R)"            
#>  [55] "SMP(L)"            "LAL(R)"            "AMMC"             
#>  [58] "ATL(R)"            "INP"               "a'3(R)"           
#>  [61] "WED(R)"            "SNP(L)"            "FBl8"             
#>  [64] "ROB(R)"            "PB(L6)"            "MB(+ACA)(R)"      
#>  [67] "FBl2"              "FBl3"              "BU(L)"            
#>  [70] "NO1(R)"            "EB"                "GNG"              
#>  [73] "b'1(R)"            "AVLP(R)"           "RUB(R)"           
#>  [76] "ATL(L)"            "b'L(R)"            "PB(R9)"           
#>  [79] "PVLP(R)"           "EPA(L)"            "b'L(L)"           
#>  [82] "SCL(L)"            "NO1(L)"            "PB(R8)"           
#>  [85] "VMNP"              "PB(R4)"            "NO(L)"            
#>  [88] "gL(R)"             "FBl1"              "FBl6"             
#>  [91] "CA(R)"             "NO2(R)"            "OL(R)"            
#>  [94] "a2(R)"             "LAL(-GA)(R)"       "NO3(R)"           
#>  [97] "ICL(R)"            "CAN(R)"            "aL(L)"            
#> [100] "SMP(R)"            "b'2(R)"            "EBr1"             
#> [103] "PB(R2)"            "SLP(R)"            "CX"               
#> [106] "PB(R7)"            "g1(R)"             "EBr6"             
#> [109] "gL(L)"             "EBr3pw"            "mALT(L)"          
#> [112] "PB(L2)"            "PRW"               "lACA(R)"          
#> [115] "a'2(R)"            "CRE(-ROB,-RUB)(R)" "PB(R6)"           
#> [118] "GOR(L)"            "PB(R5)"            "CRE(R)"           
#> [121] "EBr2r4"            "PB(L3)"            "GA(R)"            
#> [124] "aL(R)"             "LOP(R)"            "g3(R)"            
#> [127] "FB"                "g5(R)"             "CRE(L)"           
#> [130] "VES(R)"            "LX(L)"             "a'L(L)"           
#> [133] "PED(R)"            "EBr3d"             "g4(R)"            
#> [136] "b1(R)"             "PB(L1)"            "LX(R)"            
#> [139] "ME(R)"             "mALT(R)"           "NO(R)"            
#> [142] "EBr5"              "LAL(L)"            "IPS(R)"           
#> [145] "VES(L)"            "LO(R)"             "FBl5"             
#> [148] "AOT(R)"            "b2(R)"             "PB(L9)"           
#> [151] "PB(L7)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-VM1(R)"         "g2(R)"             "BU(L)"            
#>   [4] "bL(L)"             "LO(R)"             "AL-DA2(R)"        
#>   [7] "AL-DC2(R)"         "NO2(L)"            "EB"               
#>  [10] "AL-DM6(L)"         "AL-VC3m(R)"        "CRE(-ROB,-RUB)(R)"
#>  [13] "mALT(L)"           "EPA(R)"            "AL-DC1(R)"        
#>  [16] "vACA(R)"           "PB(L8)"            "AL-VC4(R)"        
#>  [19] "AL-VP3(R)"         "AMMC"              "dACA(R)"          
#>  [22] "AL-DP1m(L)"        "VLNP(R)"           "CX"               
#>  [25] "FBl4"              "AL-DL3(R)"         "AL-DP1m(R)"       
#>  [28] "EBr3am"            "WED(R)"            "PB(L4)"           
#>  [31] "a'L(L)"            "OL(R)"             "AL-VL2a(R)"       
#>  [34] "EBr2r4"            "FB-column3"        "PB(R6)"           
#>  [37] "g1(R)"             "AVLP(R)"           "LX(L)"            
#>  [40] "AL-D(L)"           "ATL(R)"            "AL-VM5v(R)"       
#>  [43] "AL-DM6(R)"         "FLA(R)"            "NO1(R)"           
#>  [46] "NO(R)"             "PB(L3)"            "b'L(R)"           
#>  [49] "AL-VC2(R)"         "b'1(R)"            "AL(L)"            
#>  [52] "PB(L9)"            "CRE(-RUB)(L)"      "g5(R)"            
#>  [55] "a3(R)"             "AL-DM5(R)"         "NO1(L)"           
#>  [58] "EBr6"              "AL-VA2(R)"         "PB(R3)"           
#>  [61] "AL-DC3(R)"         "AL-DL4(R)"         "b2(R)"            
#>  [64] "ME(R)"             "AL-DA3(L)"         "AL-VL2p(R)"       
#>  [67] "AL-VA6(L)"         "SMP(R)"            "LX(R)"            
#>  [70] "g4(R)"             "gL(R)"             "ROB(R)"           
#>  [73] "AL-VP1m(R)"        "AL-VA7m(R)"        "g3(R)"            
#>  [76] "FBl2"              "AL-VM2(R)"         "AL-VP2(R)"        
#>  [79] "a'3(R)"            "SMP(L)"            "EBr5"             
#>  [82] "a'L(R)"            "NO"                "PB(L7)"           
#>  [85] "b1(R)"             "b'L(L)"            "AL-VA5(R)"        
#>  [88] "FB"                "SCL(R)"            "IB"               
#>  [91] "MB(+ACA)(R)"       "CA(R)"             "INP"              
#>  [94] "bL(R)"             "LAL(L)"            "PED(R)"           
#>  [97] "AB(R)"             "aL(L)"             "AL-D(R)"          
#> [100] "AL-VM7d(R)"        "SAD(-AMMC)"        "AME(R)"           
#> [103] "RUB(L)"            "AL-DM3(R)"         "mALT(R)"          
#> [106] "PB(R8)"            "SIP(R)"            "AL-DL2d(R)"       
#> [109] "AL-VC1(R)"         "PB(R7)"            "aL(R)"            
#> [112] "AL-VC5(R)"         "PENP"              "b'2(R)"           
#> [115] "ICL(L)"            "AL-VM5d(R)"        "AL-VP1l(R)"       
#> [118] "AOTU(R)"           "EBr1"              "AL-DA3(R)"        
#> [121] "AL-DM1(R)"         "NO3(R)"            "EBr3d"            
#> [124] "SPS(L)"            "PB(R2)"            "FBl3"             
#> [127] "gL(L)"             "AL-DC2(L)"         "AL-VA1v(R)"       
#> [130] "AL-DL4(L)"         "AL(R)"             "POC"              
#> [133] "PB(L5)"            "ATL(L)"            "PB(R4)"           
#> [136] "AL-DC4(R)"         "AL-V(R)"           "lACA(R)"          
#> [139] "NO(L)"             "FBl1"              "AL-DL1(R)"        
#> [142] "AL-VA1d(R)"        "BU(R)"             "PB(R1)"           
#> [145] "PB(L6)"            "AL-DL5(R)"         "FBl6"             
#> [148] "AL-DM4(R)"         "AL-DM2(L)"         "AL-DM4(L)"        
#> [151] "SIP(L)"            "PVLP(R)"           "GNG"              
#> [154] "GC"                "MB(L)"             "PB(L2)"           
#> [157] "AL-VM7v(L)"        "ICL(R)"            "AL-DC4(L)"        
#> [160] "FBl7"              "LH(R)"             "hemibrain"        
#> [163] "PB(L1)"            "AL-DM3(L)"         "a'1(R)"           
#> [166] "VES(R)"            "a2(R)"             "AL-VM7d(L)"       
#> [169] "AL-VP4(R)"         "AL-VA3(R)"         "AL-VL1(R)"        
#> [172] "EPA(L)"            "AL-VA6(R)"         "CRE(L)"           
#> [175] "EBr3pw"            "AL-DM1(L)"         "PRW"              
#> [178] "AL-DL5(L)"         "NO3(L)"            "NO2(R)"           
#> [181] "AL-DA1(R)"         "AL-VM7v(R)"        "SAD"              
#> [184] "SPS(R)"            "CRE(R)"            "FBl9"             
#> [187] "AL-DP1l(R)"        "PB(R5)"            "GF(R)"            
#> [190] "GA(R)"             "AL-VM3(R)"         "AB(L)"            
#> [193] "AL-DC1(L)"         "PB(R9)"            "AL-VA4(R)"        
#> [196] "AL-VM4(R)"         "AL-VP1d(R)"        "LOP(R)"           
#> [199] "GOR(L)"            "VMNP"              "GOR(R)"           
#> [202] "MB(R)"             "AL-DA4l(R)"        "AL-VC3l(R)"       
#> [205] "a'2(R)"            "FBl8"              "a1(R)"            
#> [208] "CA(L)"             "PB"                "SNP(R)"           
#> [211] "SNP(L)"            "SCL(L)"            "AL-DA2(L)"        
#> [214] "AL-DA4m(L)"        "AOT(R)"            "AL-DM2(R)"        
#> [217] "RUB(R)"            "SLP(R)"            "AL-DL2v(R)"       
#> [220] "LAL(-GA)(R)"       "FBl5"              "LAL(R)"           
#> [223] "VES(L)"            "AL-DA4m(R)"        "AL-VA7l(R)"       
#> [226] "PLP(R)"            "AL-DM5(L)"         "AL-VP5(R)"        
#> [229] "IPS(R)"            "CAN(R)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "SAD(-AMMC)"        "FBl9"              "FBl7"             
#>   [4] "AL-VP1m(R)"        "SPS(R)"            "BU(R)"            
#>   [7] "CA(R)"             "GOR(L)"            "ME(R)"            
#>  [10] "NO(L)"             "AL-DM5(R)"         "AL-VL2p(R)"       
#>  [13] "NO3(L)"            "LO(R)"             "POC"              
#>  [16] "AL(L)"             "AL-VC2(R)"         "AL-DM2(R)"        
#>  [19] "AL-VM3(R)"         "a2(R)"             "VES(L)"           
#>  [22] "AL-VL2a(R)"        "AL-V(R)"           "PB(R2)"           
#>  [25] "AL-DA3(R)"         "AL-VA6(L)"         "PB(R1)"           
#>  [28] "AL-VC1(R)"         "PB"                "PB(L6)"           
#>  [31] "PB(R4)"            "AL-DC2(R)"         "RUB(L)"           
#>  [34] "AL-VA7l(R)"        "PENP"              "b2(R)"            
#>  [37] "b'1(R)"            "PB(L8)"            "AL-D(R)"          
#>  [40] "b1(R)"             "PB(R8)"            "PB(R3)"           
#>  [43] "AL-VA1d(R)"        "PB(L2)"            "AL-DM4(L)"        
#>  [46] "CRE(-RUB)(L)"      "SMP(R)"            "g1(R)"            
#>  [49] "SPS(L)"            "GF(R)"             "AL(R)"            
#>  [52] "FBl6"              "AL-VP1d(R)"        "SMP(L)"           
#>  [55] "PB(L7)"            "RUB(R)"            "AL-DA2(R)"        
#>  [58] "AL-VA2(R)"         "ATL(R)"            "a1(R)"            
#>  [61] "b'2(R)"            "FB-column3"        "AL-VA7m(R)"       
#>  [64] "AL-VC3l(R)"        "EBr6"              "AL-DM3(L)"        
#>  [67] "CX"                "NO2(L)"            "LH(R)"            
#>  [70] "AL-VC4(R)"         "a'1(R)"            "IPS(R)"           
#>  [73] "b'L(L)"            "AL-VM5d(R)"        "AL-VP3(R)"        
#>  [76] "AL-DA4m(L)"        "AL-DC4(L)"         "AL-DA1(R)"        
#>  [79] "AL-DC3(R)"         "AL-DM4(R)"         "AL-DP1l(R)"       
#>  [82] "AL-VL1(R)"         "EPA(R)"            "MB(R)"            
#>  [85] "AL-VP1l(R)"        "VLNP(R)"           "LX(R)"            
#>  [88] "AL-DP1m(R)"        "PB(R9)"            "MB(L)"            
#>  [91] "g2(R)"             "a3(R)"             "AL-DA4l(R)"       
#>  [94] "AL-DL1(R)"         "AL-VC3m(R)"        "SCL(R)"           
#>  [97] "EBr2r4"            "AL-D(L)"           "FBl3"             
#> [100] "CAN(R)"            "dACA(R)"           "gL(L)"            
#> [103] "VMNP"              "FBl2"              "AL-VP4(R)"        
#> [106] "vACA(R)"           "AOTU(R)"           "PB(R7)"           
#> [109] "mALT(L)"           "AL-DA3(L)"         "LX(L)"            
#> [112] "AL-DM6(L)"         "VES(R)"            "PB(L4)"           
#> [115] "AL-VC5(R)"         "a'3(R)"            "FB"               
#> [118] "FLA(R)"            "LOP(R)"            "AME(R)"           
#> [121] "AL-DL5(R)"         "PB(L5)"            "AL-VA1v(R)"       
#> [124] "AL-VM5v(R)"        "PRW"               "FBl1"             
#> [127] "SIP(R)"            "ICL(R)"            "EBr3am"           
#> [130] "FBl4"              "AB(L)"             "FBl5"             
#> [133] "AL-VA3(R)"         "hemibrain"         "gL(R)"            
#> [136] "AL-DC1(L)"         "SCL(L)"            "PB(L3)"           
#> [139] "LAL(-GA)(R)"       "AL-DM5(L)"         "AL-DL5(L)"        
#> [142] "aL(L)"             "GOR(R)"            "CA(L)"            
#> [145] "EBr5"              "g4(R)"             "AL-VM7d(L)"       
#> [148] "AL-DA2(L)"         "INP"               "g3(R)"            
#> [151] "mALT(R)"           "NO1(R)"            "PVLP(R)"          
#> [154] "PED(R)"            "GC"                "ICL(L)"           
#> [157] "AL-DL2d(R)"        "PB(R6)"            "ROB(R)"           
#> [160] "CRE(R)"            "OL(R)"             "AL-DA4m(R)"       
#> [163] "AL-DL3(R)"         "AMMC"              "SIP(L)"           
#> [166] "lACA(R)"           "SNP(R)"            "SNP(L)"           
#> [169] "NO1(L)"            "AL-VP2(R)"         "AL-DL4(R)"        
#> [172] "AL-VM7d(R)"        "AL-VM7v(R)"        "aL(R)"            
#> [175] "AB(R)"             "AL-VP5(R)"         "a'L(R)"           
#> [178] "AL-VA6(R)"         "GNG"               "AOT(R)"           
#> [181] "AL-DM1(L)"         "FBl8"              "PLP(R)"           
#> [184] "AL-DL4(L)"         "CRE(-ROB,-RUB)(R)" "SAD"              
#> [187] "EBr1"              "AVLP(R)"           "AL-DM1(R)"        
#> [190] "AL-VM2(R)"         "LAL(L)"            "EB"               
#> [193] "CRE(L)"            "LAL(R)"            "AL-VA5(R)"        
#> [196] "AL-DC2(L)"         "NO(R)"             "a'L(L)"           
#> [199] "AL-DM3(R)"         "NO"                "NO2(R)"           
#> [202] "EBr3d"             "AL-DM2(L)"         "AL-DP1m(L)"       
#> [205] "EPA(L)"            "AL-DL2v(R)"        "AL-DM6(R)"        
#> [208] "NO3(R)"            "BU(L)"             "a'2(R)"           
#> [211] "PB(L9)"            "bL(L)"             "GA(R)"            
#> [214] "PB(L1)"            "AL-VM7v(L)"        "EBr3pw"           
#> [217] "MB(+ACA)(R)"       "SLP(R)"            "AL-VM1(R)"        
#> [220] "WED(R)"            "b'L(R)"            "ATL(L)"           
#> [223] "AL-DC4(R)"         "IB"                "PB(R5)"           
#> [226] "bL(R)"             "AL-VA4(R)"         "AL-VM4(R)"        
#> [229] "g5(R)"             "AL-DC1(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`manc:v1.0`
#> $`manc:v1.0`$`last-mod`
#> [1] "2023-05-31 23:54:01"
#> 
#> $`manc:v1.0`$uuid
#> [1] "59b37970bc7a4341b9a3a965a0d6b402"
#> 
#> $`manc:v1.0`$ROIs
#>  [1] "ProCN(R)"         "Ov(R)"            "WTct(UTct-T2)(L)" "IntTct"          
#>  [5] "AbN3(L)"          "CvN(R)"           "DMetaN(L)"        "PDMN(R)"         
#>  [9] "ProLN(L)"         "mVAC(T1)(R)"      "LegNp(T2)(L)"     "NTct(UTct-T1)(R)"
#> [13] "ADMN(R)"          "AbN1(R)"          "LegNp(T1)(R)"     "AbNT"            
#> [17] "MesoAN(L)"        "LegNp(T2)(R)"     "HTct(UTct-T3)(L)" "AbN1(L)"         
#> [21] "MetaLN(R)"        "ProAN(R)"         "CV"               "MesoLN(R)"       
#> [25] "PDMN(L)"          "NTct(UTct-T1)(L)" "mVAC(T2)(L)"      "mVAC(T2)(R)"     
#> [29] "ANm"              "DProN(L)"         "MesoAN(R)"        "PrN(R)"          
#> [33] "LegNp(T3)(R)"     "AbN4(R)"          "DProN(R)"         "MetaLN(L)"       
#> [37] "PrN(L)"           "AbN2(L)"          "MesoLN(L)"        "WTct(UTct-T2)(R)"
#> [41] "VProN(L)"         "VProN(R)"         "ADMN(L)"          "GF(L)"           
#> [45] "mVAC(T3)(R)"      "mVAC(T1)(L)"      "mVAC(T3)(L)"      "LTct"            
#> [49] "AbN2(R)"          "DMetaN(R)"        "LegNp(T1)(L)"     "CvN(L)"          
#> [53] "LegNp(T3)(L)"     "GF(R)"            "AbN3(R)"          "AbN4(L)"         
#> [57] "ProLN(R)"         "ProCN(L)"         "Ov(L)"            "ProAN(L)"        
#> [61] "HTct(UTct-T3)(R)"
#> 
#> $`manc:v1.0`$superLevelROIs
#>  [1] "ADMN(L)"          "ADMN(R)"          "Ov(L)"            "Ov(R)"           
#>  [5] "ANm"              "AbN1(L)"          "AbN1(R)"          "AbN2(L)"         
#>  [9] "AbN2(R)"          "AbN3(L)"          "AbN3(R)"          "AbN4(L)"         
#> [13] "AbN4(R)"          "AbNT"             "CV"               "CvN(L)"          
#> [17] "CvN(R)"           "DMetaN(L)"        "DMetaN(R)"        "DProN(L)"        
#> [21] "DProN(R)"         "GF(L)"            "GF(R)"            "HTct(UTct-T3)(L)"
#> [25] "HTct(UTct-T3)(R)" "LegNp(T1)(L)"     "LegNp(T1)(R)"     "LegNp(T2)(L)"    
#> [29] "LegNp(T2)(R)"     "LegNp(T3)(L)"     "LegNp(T3)(R)"     "IntTct"          
#> [33] "LTct"             "MesoAN(L)"        "MesoAN(R)"        "MesoLN(L)"       
#> [37] "MesoLN(R)"        "MetaLN(L)"        "MetaLN(R)"        "NTct(UTct-T1)(L)"
#> [41] "NTct(UTct-T1)(R)" "PDMN(L)"          "PrN(L)"           "PrN(R)"          
#> [45] "ProCN(L)"         "ProCN(R)"         "PDMN(R)"          "ProAN(L)"        
#> [49] "ProAN(R)"         "ProLN(L)"         "ProLN(R)"         "VProN(L)"        
#> [53] "VProN(R)"         "WTct(UTct-T2)(L)" "WTct(UTct-T2)(R)" "mVAC(T1)(L)"     
#> [57] "mVAC(T1)(R)"      "mVAC(T2)(L)"      "mVAC(T2)(R)"      "mVAC(T3)(L)"     
#> [61] "mVAC(T3)(R)"     
#> 
#> $`manc:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/manc-connectome"
#> 
#> $`manc:v1.0`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method    path                                    
#> [1,] "OPTIONS" "/api/*"                                
#> [2,] "POST"    "/api/raw/cypher/transaction/:id/cypher"
#> [3,] "GET"     "/api/available"                        
#> [4,] "REPORT"  "/api"                                  
#> [5,] "GET"     "/api/npexplorer/nglayers*"             
#> [6,] "GET"     "/api/cached/roicompleteness"           
#>      name                                                                                    
#> [1,] "github.com/labstack/echo/v4.glob..func1"                                               
#> [2,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.SetupRoutes.func1"                     
#> [4,] "github.com/labstack/echo/v4.glob..func1"                                               
#> [5,] "github.com/labstack/echo/v4.StaticDirectoryHandler.func1"                              
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROICompleteness-fm"
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }