Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0`
#> $`hemibrain:v1.0`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0`$ROIs
#>   [1] "SNP(L)"            "PVLP(R)"           "PB(L3)"           
#>   [4] "WED(R)"            "ME(R)"             "EBr2r4"           
#>   [7] "PB(R9)"            "FBl3"              "aL(L)"            
#>  [10] "SMP(R)"            "hemibrain"         "SPS(R)"           
#>  [13] "SCL(R)"            "LX(R)"             "ICL(L)"           
#>  [16] "EBr3am"            "mALT(L)"           "a1(R)"            
#>  [19] "lACA(R)"           "NO1(R)"            "PB(R2)"           
#>  [22] "CAN(R)"            "PB(R4)"            "FB"               
#>  [25] "EB"                "SAD(-AMMC)"        "VES(R)"           
#>  [28] "AME(R)"            "NO1(L)"            "GF(R)"            
#>  [31] "VES(L)"            "gL(L)"             "LAL(R)"           
#>  [34] "SCL(L)"            "PB(L5)"            "GOR(L)"           
#>  [37] "NO3(L)"            "FBl6"              "EBr5"             
#>  [40] "PLP(R)"            "BU(R)"             "PB(R5)"           
#>  [43] "EPA(R)"            "SIP(R)"            "ATL(R)"           
#>  [46] "PB(L4)"            "PB(R6)"            "LAL(L)"           
#>  [49] "PB(R7)"            "b1(R)"             "b'L(L)"           
#>  [52] "ATL(L)"            "AOTU(R)"           "dACA(R)"          
#>  [55] "SNP(R)"            "g4(R)"             "a'1(R)"           
#>  [58] "CRE(R)"            "ICL(R)"            "AL(R)"            
#>  [61] "PB(L2)"            "a'3(R)"            "FBl9"             
#>  [64] "IPS(R)"            "SLP(R)"            "SAD"              
#>  [67] "GC"                "a'L(R)"            "PB(R1)"           
#>  [70] "FBl5"              "CX"                "CRE(-ROB,-RUB)(R)"
#>  [73] "b'2(R)"            "g5(R)"             "FBl2"             
#>  [76] "AVLP(R)"           "AOT(R)"            "AMMC"             
#>  [79] "g3(R)"             "PB"                "EBr3pw"           
#>  [82] "FBl8"              "a3(R)"             "LH(R)"            
#>  [85] "b'1(R)"            "PED(R)"            "b2(R)"            
#>  [88] "EBr1"              "NO"                "PB(L9)"           
#>  [91] "mALT(R)"           "PB(L6)"            "NO(R)"            
#>  [94] "CRE(L)"            "INP"               "PENP"             
#>  [97] "NO(L)"             "LOP(R)"            "LO(R)"            
#> [100] "PRW"               "PB(L1)"            "a'2(R)"           
#> [103] "bL(L)"             "FBl1"              "AB(L)"            
#> [106] "GA(R)"             "aL(R)"             "CA(R)"            
#> [109] "GNG"               "MB(R)"             "SMP(L)"           
#> [112] "a2(R)"             "SIP(L)"            "VMNP"             
#> [115] "bL(R)"             "IB"                "LAL(-GA)(R)"      
#> [118] "PB(R3)"            "EPA(L)"            "CA(L)"            
#> [121] "POC"               "vACA(R)"           "g1(R)"            
#> [124] "g2(R)"             "PB(L8)"            "MB(+ACA)(R)"      
#> [127] "NO2(R)"            "a'L(L)"            "NO3(R)"           
#> [130] "MB(L)"             "EBr6"              "SPS(L)"           
#> [133] "PB(R8)"            "FBl4"              "LX(L)"            
#> [136] "BU(L)"             "ROB(R)"            "FBl7"             
#> [139] "RUB(L)"            "b'L(R)"            "PB(L7)"           
#> [142] "AB(R)"             "GOR(R)"            "RUB(R)"           
#> [145] "EBr3d"             "gL(R)"             "NO2(L)"           
#> [148] "FLA(R)"            "AL(L)"             "OL(R)"            
#> [151] "VLNP(R)"          
#> 
#> $`hemibrain:v1.0`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method     path                                
#> [1,] "GET"      "/api/dbmeta/database"              
#> [2,] "GET"      "/api/v:ver/npexplorer/completeness"
#> [3,] "PROPFIND" "/api/*"                            
#> [4,] "GET"      "/*"                                
#> [5,] "POST"     "/api/npexplorer/roisinneuron"      
#> [6,] "POST"     "/api/npexplorer/autapses"          
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatabase-fm"            
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [3,] "github.com/labstack/echo.glob..func1"                                                      
#> [4,] "github.com/labstack/echo.common.static.func1"                                              
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"    

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }