Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB"                "ICL(R)"            "PB(L5)"           
#>   [4] "mALT(R)"           "EBr1"              "FBl1"             
#>   [7] "FBl4"              "a'3(R)"            "GF(R)"            
#>  [10] "PB(L8)"            "FBl2"              "NO2(R)"           
#>  [13] "GOR(R)"            "OL(R)"             "GC"               
#>  [16] "PB(L6)"            "NO3(L)"            "SNP(R)"           
#>  [19] "INP"               "VLNP(R)"           "AB(R)"            
#>  [22] "EBr6"              "g5(R)"             "LO(R)"            
#>  [25] "SCL(R)"            "g1(R)"             "FB"               
#>  [28] "bL(L)"             "FBl8"              "RUB(R)"           
#>  [31] "mALT(L)"           "PED(R)"            "PB(R5)"           
#>  [34] "AL(R)"             "a'1(R)"            "FBl9"             
#>  [37] "EPA(L)"            "EBr5"              "SAD(-AMMC)"       
#>  [40] "aL(L)"             "MB(R)"             "AB(L)"            
#>  [43] "AMMC"              "FLA(R)"            "PB(R1)"           
#>  [46] "PB(R6)"            "VMNP"              "PB(R9)"           
#>  [49] "FBl3"              "CX"                "a1(R)"            
#>  [52] "MB(L)"             "dACA(R)"           "MB(+ACA)(R)"      
#>  [55] "LAL(R)"            "PB(L4)"            "a'L(L)"           
#>  [58] "GA(R)"             "AOTU(R)"           "CRE(R)"           
#>  [61] "g4(R)"             "BU(R)"             "FBl7"             
#>  [64] "CA(R)"             "LOP(R)"            "g3(R)"            
#>  [67] "CA(L)"             "ROB(R)"            "b'2(R)"           
#>  [70] "hemibrain"         "PB(L1)"            "EBr3d"            
#>  [73] "EBr2r4"            "ATL(R)"            "EBr3am"           
#>  [76] "a2(R)"             "ICL(L)"            "LX(L)"            
#>  [79] "PB(R4)"            "b'L(L)"            "IB"               
#>  [82] "AL(L)"             "PENP"              "GOR(L)"           
#>  [85] "CRE(-ROB,-RUB)(R)" "SIP(L)"            "SLP(R)"           
#>  [88] "FBl6"              "POC"               "SMP(R)"           
#>  [91] "vACA(R)"           "WED(R)"            "NO1(L)"           
#>  [94] "b'L(R)"            "EB"                "IPS(R)"           
#>  [97] "CAN(R)"            "GNG"               "NO(L)"            
#> [100] "SMP(L)"            "EPA(R)"            "PB(L9)"           
#> [103] "FBl5"              "b1(R)"             "ATL(L)"           
#> [106] "a'L(R)"            "PRW"               "PB(R7)"           
#> [109] "g2(R)"             "SPS(L)"            "gL(L)"            
#> [112] "SAD"               "PVLP(R)"           "PB(R3)"           
#> [115] "PB(L2)"            "CRE(L)"            "SPS(R)"           
#> [118] "SCL(L)"            "PB(L3)"            "AVLP(R)"          
#> [121] "AOT(R)"            "LAL(-GA)(R)"       "aL(R)"            
#> [124] "a'2(R)"            "EBr3pw"            "SNP(L)"           
#> [127] "VES(R)"            "LAL(L)"            "SIP(R)"           
#> [130] "PB(R2)"            "bL(R)"             "a3(R)"            
#> [133] "NO1(R)"            "NO"                "VES(L)"           
#> [136] "b2(R)"             "BU(L)"             "PB(R8)"           
#> [139] "PLP(R)"            "LX(R)"             "ME(R)"            
#> [142] "gL(R)"             "PB(L7)"            "NO3(R)"           
#> [145] "LH(R)"             "NO(R)"             "AME(R)"           
#> [148] "RUB(L)"            "b'1(R)"            "NO2(L)"           
#> [151] "lACA(R)"          
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "PRW"               "FB"                "SNP(R)"           
#>   [4] "FBl5"              "AL-DA3(R)"         "mALT(L)"          
#>   [7] "AL-DL4(R)"         "vACA(R)"           "EBr3pw"           
#>  [10] "AL(R)"             "AL-DL4(L)"         "NO3(L)"           
#>  [13] "ICL(L)"            "FBl9"              "IPS(R)"           
#>  [16] "b'L(R)"            "SPS(L)"            "g4(R)"            
#>  [19] "AL-DM5(R)"         "PB(R8)"            "LX(L)"            
#>  [22] "AL-VP1d(R)"        "VLNP(R)"           "a2(R)"            
#>  [25] "PB(L1)"            "FBl1"              "AL-DM1(R)"        
#>  [28] "AL-VA2(R)"         "ATL(L)"            "PB(R7)"           
#>  [31] "bL(L)"             "AL-DL1(R)"         "AL-VP3(R)"        
#>  [34] "SAD"               "NO(R)"             "PVLP(R)"          
#>  [37] "AL-DL5(R)"         "a'L(R)"            "AL-DM4(L)"        
#>  [40] "PB(R5)"            "CA(L)"             "AL-D(L)"          
#>  [43] "LO(R)"             "AL-DP1m(R)"        "EBr6"             
#>  [46] "AL-DA2(R)"         "AMMC"              "LAL(-GA)(R)"      
#>  [49] "a'L(L)"            "AL-DA4l(R)"        "AL-VM7d(R)"       
#>  [52] "b2(R)"             "BU(R)"             "SNP(L)"           
#>  [55] "FBl6"              "AL-DC1(R)"         "AL-DM6(R)"        
#>  [58] "VES(R)"            "AL-DA1(R)"         "AL-DL2v(R)"       
#>  [61] "b'2(R)"            "AL-DM1(L)"         "AL-DM3(L)"        
#>  [64] "PB(L6)"            "AL-VP1l(R)"        "hemibrain"        
#>  [67] "g1(R)"             "dACA(R)"           "gL(L)"            
#>  [70] "AL-DL5(L)"         "AL-VA7m(R)"        "AL-VC3m(R)"       
#>  [73] "g3(R)"             "LX(R)"             "PB(L7)"           
#>  [76] "MB(R)"             "NO1(R)"            "MB(+ACA)(R)"      
#>  [79] "ICL(R)"            "AL-VL1(R)"         "AL-VA3(R)"        
#>  [82] "AL-V(R)"           "AL-VP1m(R)"        "AL-VA6(R)"        
#>  [85] "PB(L3)"            "AL-VL2a(R)"        "AL-VL2p(R)"       
#>  [88] "IB"                "SCL(R)"            "EBr5"             
#>  [91] "AL-DM5(L)"         "NO2(R)"            "SLP(R)"           
#>  [94] "AL-DC2(R)"         "AL-VC2(R)"         "lACA(R)"          
#>  [97] "g2(R)"             "mALT(R)"           "BU(L)"            
#> [100] "LAL(L)"            "MB(L)"             "g5(R)"            
#> [103] "aL(L)"             "AL-DL2d(R)"        "NO"               
#> [106] "gL(R)"             "AL-DM3(R)"         "AL-VC5(R)"        
#> [109] "EBr2r4"            "NO2(L)"            "PB(R6)"           
#> [112] "CA(R)"             "EPA(L)"            "OL(R)"            
#> [115] "AL-VM4(R)"         "b1(R)"             "GNG"              
#> [118] "b'L(L)"            "SMP(R)"            "b'1(R)"           
#> [121] "WED(R)"            "FBl3"              "EBr3am"           
#> [124] "RUB(R)"            "AL-DC2(L)"         "AL-DP1m(L)"       
#> [127] "a'1(R)"            "ATL(R)"            "AVLP(R)"          
#> [130] "AL-DP1l(R)"        "AL-VM5d(R)"        "EBr3d"            
#> [133] "AL-VA6(L)"         "FBl8"              "AL-DL3(R)"        
#> [136] "AOT(R)"            "AB(L)"             "a'3(R)"           
#> [139] "AL-VA4(R)"         "EB"                "LH(R)"            
#> [142] "AL-DA4m(R)"        "PENP"              "VMNP"             
#> [145] "AB(R)"             "AL-VA7l(R)"        "SCL(L)"           
#> [148] "NO1(L)"            "NO(L)"             "SMP(L)"           
#> [151] "AL-DC4(L)"         "SIP(R)"            "CAN(R)"           
#> [154] "AL-VC1(R)"         "AOTU(R)"           "AL-DA2(L)"        
#> [157] "AL-DA3(L)"         "AL-DC1(L)"         "AL-VM7d(L)"       
#> [160] "FBl2"              "PB(L4)"            "AL-VM1(R)"        
#> [163] "CRE(R)"            "CRE(-RUB)(L)"      "FLA(R)"           
#> [166] "EPA(R)"            "PB(R4)"            "AL-VA1d(R)"       
#> [169] "AL-VA5(R)"         "CX"                "GA(R)"            
#> [172] "EBr1"              "INP"               "AL-VC3l(R)"       
#> [175] "LOP(R)"            "SPS(R)"            "NO3(R)"           
#> [178] "AL-VA1v(R)"        "AL-VP4(R)"         "aL(R)"            
#> [181] "ME(R)"             "RUB(L)"            "a3(R)"            
#> [184] "AL-DM2(R)"         "AL-D(R)"           "GC"               
#> [187] "AL-DA4m(L)"        "ROB(R)"            "AL-VP2(R)"        
#> [190] "PLP(R)"            "a'2(R)"            "PB(R1)"           
#> [193] "CRE(L)"            "SAD(-AMMC)"        "AL-VM2(R)"        
#> [196] "GOR(L)"            "AL(L)"             "AL-DM2(L)"        
#> [199] "AL-VM5v(R)"        "PB(L8)"            "FBl4"             
#> [202] "FBl7"              "AL-DM6(L)"         "PB(R9)"           
#> [205] "PB(R3)"            "AL-DC3(R)"         "PB(L5)"           
#> [208] "bL(R)"             "LAL(R)"            "PB"               
#> [211] "FB-column3"        "VES(L)"            "AL-VC4(R)"        
#> [214] "AL-VM7v(L)"        "PB(L9)"            "PED(R)"           
#> [217] "AME(R)"            "GOR(R)"            "AL-DC4(R)"        
#> [220] "AL-VP5(R)"         "PB(L2)"            "a1(R)"            
#> [223] "GF(R)"             "SIP(L)"            "PB(R2)"           
#> [226] "AL-DM4(R)"         "AL-VM7v(R)"        "POC"              
#> [229] "CRE(-ROB,-RUB)(R)" "AL-VM3(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.2`
#> $`hemibrain:v1.2`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2`$ROIs
#>   [1] "ATL(L)"            "EBr3am"            "FB-column3"       
#>   [4] "SLP(R)"            "AL-DM5(R)"         "AL-D(R)"          
#>   [7] "AL-VL1(R)"         "AL-DA2(L)"         "AL-DA4l(R)"       
#>  [10] "AL-VA7l(R)"        "SCL(L)"            "AMMC"             
#>  [13] "NO3(R)"            "EBr3d"             "EBr6"             
#>  [16] "AL-DA3(L)"         "bL(L)"             "AL-VA1d(R)"       
#>  [19] "NO1(L)"            "AL-D(L)"           "NO"               
#>  [22] "MB(R)"             "AL-DP1l(R)"        "PB(R1)"           
#>  [25] "AVLP(R)"           "AOT(R)"            "SAD"              
#>  [28] "a'1(R)"            "b1(R)"             "AL-DL5(R)"        
#>  [31] "AL-VM5v(R)"        "g2(R)"             "AL-DL4(L)"        
#>  [34] "AL-VM1(R)"         "CA(L)"             "AL-DM6(L)"        
#>  [37] "PB(L9)"            "SNP(L)"            "MB(+ACA)(R)"      
#>  [40] "FBl8"              "g5(R)"             "LH(R)"            
#>  [43] "AL-DA3(R)"         "AL-DA4m(R)"        "AL-VP3(R)"        
#>  [46] "aL(L)"             "b'2(R)"            "CRE(-RUB)(L)"     
#>  [49] "FBl7"              "AL-DC3(R)"         "mALT(R)"          
#>  [52] "gL(R)"             "FBl4"              "EPA(L)"           
#>  [55] "VES(L)"            "b2(R)"             "g4(R)"            
#>  [58] "AL-VC3m(R)"        "AB(L)"             "SPS(R)"           
#>  [61] "AL-DA1(R)"         "NO2(L)"            "AL-VM7d(R)"       
#>  [64] "b'1(R)"            "PB(L2)"            "g1(R)"            
#>  [67] "dACA(R)"           "AL-DL5(L)"         "AL-DM3(L)"        
#>  [70] "AL-DC2(R)"         "CX"                "a2(R)"            
#>  [73] "GOR(L)"            "AL-DP1m(L)"        "PB(R3)"           
#>  [76] "PB(R5)"            "EBr1"              "BU(R)"            
#>  [79] "FBl3"              "b'L(L)"            "hemibrain"        
#>  [82] "PRW"               "PB(R2)"            "AL-DA2(R)"        
#>  [85] "AL-VP1m(R)"        "AL-DC4(R)"         "AOTU(R)"          
#>  [88] "CRE(R)"            "PB(R9)"            "LAL(L)"           
#>  [91] "AL-DM4(L)"         "GNG"               "AL-VM5d(R)"       
#>  [94] "AL-VP1d(R)"        "PB(L1)"            "PB(R8)"           
#>  [97] "SAD(-AMMC)"        "a'L(L)"            "AL-VC2(R)"        
#> [100] "AL-DC4(L)"         "AL-VA6(L)"         "AL-VP1l(R)"       
#> [103] "AL-DM2(R)"         "EBr2r4"            "SIP(R)"           
#> [106] "ICL(L)"            "FBl6"              "GOR(R)"           
#> [109] "SNP(R)"            "AL-DL2v(R)"        "AL-VM7v(R)"       
#> [112] "mALT(L)"           "PED(R)"            "FBl5"             
#> [115] "PENP"              "AL-DC1(R)"         "AL-VC5(R)"        
#> [118] "gL(L)"             "AL-VC4(R)"         "AL-VM3(R)"        
#> [121] "PB(L8)"            "AL-VC1(R)"         "CRE(-ROB,-RUB)(R)"
#> [124] "NO3(L)"            "FBl2"              "AL-DM3(R)"        
#> [127] "AL-DM6(R)"         "VLNP(R)"           "PVLP(R)"          
#> [130] "ICL(R)"            "LX(L)"             "a'3(R)"           
#> [133] "FBl9"              "AL-VA3(R)"         "FB"               
#> [136] "EPA(R)"            "OL(R)"             "a'2(R)"           
#> [139] "a'L(R)"            "FBl1"              "SIP(L)"           
#> [142] "AL-VL2p(R)"        "b'L(R)"            "AL-DM4(R)"        
#> [145] "GC"                "AL-VM7v(L)"        "AL-DL1(R)"        
#> [148] "PB(L5)"            "a1(R)"             "PB(L4)"           
#> [151] "AL-VC3l(R)"        "MB(L)"             "ME(R)"            
#> [154] "EB"                "AL-VP2(R)"         "LOP(R)"           
#> [157] "EBr5"              "NO1(R)"            "ATL(R)"           
#> [160] "bL(R)"             "AL-VP5(R)"         "lACA(R)"          
#> [163] "PB(L7)"            "VES(R)"            "RUB(L)"           
#> [166] "vACA(R)"           "CRE(L)"            "IPS(R)"           
#> [169] "AL(R)"             "AL-DC2(L)"         "AL-VA4(R)"        
#> [172] "SMP(L)"            "BU(L)"             "AL-VL2a(R)"       
#> [175] "SMP(R)"            "FLA(R)"            "SPS(L)"           
#> [178] "POC"               "AL-DP1m(R)"        "AL-VM2(R)"        
#> [181] "AL-DC1(L)"         "AL-DM5(L)"         "AL-VM7d(L)"       
#> [184] "PB(L6)"            "CAN(R)"            "GF(R)"            
#> [187] "RUB(R)"            "PB(L3)"            "AL-DA4m(L)"       
#> [190] "LO(R)"             "CA(R)"             "LAL(R)"           
#> [193] "WED(R)"            "NO(L)"             "AL-VM4(R)"        
#> [196] "PB(R6)"            "AL(L)"             "AL-DL2d(R)"       
#> [199] "AL-VA6(R)"         "LAL(-GA)(R)"       "AB(R)"            
#> [202] "AL-DL3(R)"         "AL-V(R)"           "GA(R)"            
#> [205] "LX(R)"             "NO(R)"             "AL-VP4(R)"        
#> [208] "AL-VA5(R)"         "g3(R)"             "VMNP"             
#> [211] "AME(R)"            "NO2(R)"            "PB(R4)"           
#> [214] "EBr3pw"            "AL-DL4(R)"         "AL-VA2(R)"        
#> [217] "PB(R7)"            "PLP(R)"            "INP"              
#> [220] "AL-DM1(R)"         "AL-VA1v(R)"        "AL-DM1(L)"        
#> [223] "AL-DM2(L)"         "ROB(R)"            "a3(R)"            
#> [226] "aL(R)"             "SCL(R)"            "PB"               
#> [229] "AL-VA7m(R)"        "IB"               
#> 
#> $`hemibrain:v1.2`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                
#> [1,] "GET"  "/profile"                          
#> [2,] "GET"  "/api/dbmeta/datasets"              
#> [3,] "GET"  "/api/npexplorer/neuronmetavals"    
#> [4,] "GET"  "/api/npexplorer/roiconnectivity"   
#> [5,] "GET"  "/api/v:ver/npexplorer/distribution"
#> [6,] "POST" "/api/v:ver/npexplorer/completeness"
#>      name                                                                                           
#> [1,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                       
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatasets-fm"               
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMetaVals-fm" 
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIConnectivity-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }