Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DA3(L)"         "PB(R3)"            "a'L(R)"           
#>   [4] "AL-DA4m(R)"        "AL-DL2d(R)"        "PB(L2)"           
#>   [7] "AOTU(R)"           "FBl4"              "FBl5"             
#>  [10] "FBl9"              "AL-DA3(R)"         "AL-VC3m(R)"       
#>  [13] "AL-DM4(L)"         "AL-DM5(L)"         "ICL(L)"           
#>  [16] "SAD(-AMMC)"        "AL-DC2(R)"         "AL-DA1(R)"        
#>  [19] "AL-DL3(R)"         "AL-DP1l(R)"        "AL-VC2(R)"        
#>  [22] "dACA(R)"           "PB(L4)"            "AL-DC4(R)"        
#>  [25] "AL(R)"             "AL-VP4(R)"         "b'2(R)"           
#>  [28] "NO1(L)"            "AL-DC1(L)"         "VES(R)"           
#>  [31] "bL(L)"             "b2(R)"             "EBr6"             
#>  [34] "EB"                "AL-DM2(L)"         "AL-VM7v(R)"       
#>  [37] "AL-VP2(R)"         "RUB(R)"            "vACA(R)"          
#>  [40] "AL-VM3(R)"         "AL-VM5v(R)"        "AL-V(R)"          
#>  [43] "AOT(R)"            "AL-DC2(L)"         "AL-VA6(L)"        
#>  [46] "OL(R)"             "AL-VL2p(R)"        "SMP(L)"           
#>  [49] "AL-DL4(R)"         "AL-VC4(R)"         "aL(R)"            
#>  [52] "AL-DP1m(L)"        "AL-VA4(R)"         "GA(R)"            
#>  [55] "ATL(R)"            "AL-VP5(R)"         "b'1(R)"           
#>  [58] "LX(R)"             "lACA(R)"           "b'L(R)"           
#>  [61] "aL(L)"             "AL-VM7d(R)"        "a'1(R)"           
#>  [64] "ICL(R)"            "AL-VA6(R)"         "hemibrain"        
#>  [67] "EBr3pw"            "LH(R)"             "LOP(R)"           
#>  [70] "AL-DM1(L)"         "CA(R)"             "PB(L5)"           
#>  [73] "SIP(L)"            "g4(R)"             "AME(R)"           
#>  [76] "AL-DM5(R)"         "AL-VP1l(R)"        "EBr3am"           
#>  [79] "mALT(R)"           "AL-DL5(L)"         "FBl6"             
#>  [82] "PB(R4)"            "AL-VM4(R)"         "AL-VM5d(R)"       
#>  [85] "NO2(L)"            "SAD"               "FB"               
#>  [88] "ROB(R)"            "PB(L6)"            "LAL(L)"           
#>  [91] "g1(R)"             "AL-VM1(R)"         "SPS(L)"           
#>  [94] "AL-DL5(R)"         "AL-VA7m(R)"        "CRE(-ROB,-RUB)(R)"
#>  [97] "ME(R)"             "AL-DM1(R)"         "SCL(L)"           
#> [100] "FBl1"              "GOR(L)"            "VMNP"             
#> [103] "a'2(R)"            "SIP(R)"            "AL(L)"            
#> [106] "AL-VA2(R)"         "VLNP(R)"           "AVLP(R)"          
#> [109] "AL-DC3(R)"         "PED(R)"            "g2(R)"            
#> [112] "AL-D(L)"           "AL-VM7v(L)"        "EPA(R)"           
#> [115] "PRW"               "SCL(R)"            "PB(R8)"           
#> [118] "EBr2r4"            "AL-DA2(L)"         "AL-VP3(R)"        
#> [121] "NO3(R)"            "RUB(L)"            "AL-VA7l(R)"       
#> [124] "IB"                "PB(R7)"            "LAL(-GA)(R)"      
#> [127] "FBl7"              "AL-D(R)"           "gL(R)"            
#> [130] "SMP(R)"            "AL-DM3(L)"         "PB(R9)"           
#> [133] "PB(R1)"            "LO(R)"             "BU(R)"            
#> [136] "SLP(R)"            "AL-DP1m(R)"        "a2(R)"            
#> [139] "AL-VC1(R)"         "AL-VP1d(R)"        "mALT(L)"          
#> [142] "bL(R)"             "AL-VP1m(R)"        "AL-DL2v(R)"       
#> [145] "AB(L)"             "BU(L)"             "AL-DM4(R)"        
#> [148] "AL-DM6(R)"         "FB-column3"        "POC"              
#> [151] "AL-VA5(R)"         "CA(L)"             "PENP"             
#> [154] "PB(L9)"            "PB(R2)"            "AL-DA4l(R)"       
#> [157] "AL-DA4m(L)"        "VES(L)"            "AL-VA1d(R)"       
#> [160] "SPS(R)"            "PB(L7)"            "AL-VC5(R)"        
#> [163] "a'L(L)"            "MB(R)"             "AL-DC1(R)"        
#> [166] "AL-VA1v(R)"        "g3(R)"             "AL-DC4(L)"        
#> [169] "AL-DM6(L)"         "EPA(L)"            "CRE(L)"           
#> [172] "INP"               "a3(R)"             "AL-DM2(R)"        
#> [175] "PB(L1)"            "gL(L)"             "IB112"            
#> [178] "AL-VA3(R)"         "AL-DL4(L)"         "PB(L8)"           
#> [181] "GOR(R)"            "AL-VM2(R)"         "a1(R)"            
#> [184] "GC"                "FLA(R)"            "AL-VM7d(L)"       
#> [187] "NO2(R)"            "EBr1"              "CRE(-RUB)(L)"     
#> [190] "MB(+ACA)(R)"       "FBl2"              "NO3(L)"           
#> [193] "CAN(R)"            "LAL(R)"            "GNG"              
#> [196] "AMMC"              "GF(R)"             "PB"               
#> [199] "NO(R)"             "PVLP(R)"           "LX(L)"            
#> [202] "AL-DL1(R)"         "PB(R5)"            "CX"               
#> [205] "MB(L)"             "EBr5"              "g5(R)"            
#> [208] "AB(R)"             "FBl3"              "AL-DM3(R)"        
#> [211] "AL-VC3l(R)"        "ATL(L)"            "NO1(R)"           
#> [214] "SNP(L)"            "FBl8"              "b'L(L)"           
#> [217] "WED(R)"            "CRE(R)"            "SNP(R)"           
#> [220] "PB(R6)"            "EBr3d"             "AL-VL2a(R)"       
#> [223] "PLP(R)"            "NO"                "b1(R)"            
#> [226] "AL-DA2(R)"         "PB(L3)"            "IPS(R)"           
#> [229] "AL-VL1(R)"         "a'3(R)"            "NO(L)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(R6)"            "EBr1"              "EBr5"             
#>   [4] "SIP(R)"            "b'L(R)"            "FBl4"             
#>   [7] "bL(R)"             "PB(L3)"            "EBr6"             
#>  [10] "VES(R)"            "INP"               "PRW"              
#>  [13] "NO(R)"             "MB(R)"             "ATL(L)"           
#>  [16] "PB"                "EBr3d"             "ICL(R)"           
#>  [19] "AME(R)"            "g1(R)"             "aL(R)"            
#>  [22] "ICL(L)"            "GNG"               "GOR(R)"           
#>  [25] "AB(L)"             "SIP(L)"            "ATL(R)"           
#>  [28] "LX(L)"             "CAN(R)"            "FBl1"             
#>  [31] "FBl5"              "RUB(L)"            "AL(R)"            
#>  [34] "SAD"               "VMNP"              "SNP(R)"           
#>  [37] "g4(R)"             "SMP(R)"            "PB(L2)"           
#>  [40] "LOP(R)"            "EBr3pw"            "AB(R)"            
#>  [43] "EBr3am"            "LAL(-GA)(R)"       "NO3(R)"           
#>  [46] "b2(R)"             "PB(R1)"            "POC"              
#>  [49] "PB(R7)"            "PVLP(R)"           "NO2(R)"           
#>  [52] "SLP(R)"            "WED(R)"            "PB(L9)"           
#>  [55] "a'1(R)"            "FBl2"              "FBl3"             
#>  [58] "FBl8"              "GF(R)"             "LX(R)"            
#>  [61] "NO"                "g5(R)"             "lACA(R)"          
#>  [64] "b1(R)"             "EPA(L)"            "NO1(L)"           
#>  [67] "SPS(L)"            "gL(L)"             "MB(+ACA)(R)"      
#>  [70] "PB(L6)"            "RUB(R)"            "PB(L5)"           
#>  [73] "mALT(L)"           "PENP"              "hemibrain"        
#>  [76] "mALT(R)"           "FB"                "b'1(R)"           
#>  [79] "vACA(R)"           "b'2(R)"            "SPS(R)"           
#>  [82] "BU(L)"             "ME(R)"             "MB(L)"            
#>  [85] "PB(L7)"            "FBl6"              "VES(L)"           
#>  [88] "NO(L)"             "a2(R)"             "PB(L8)"           
#>  [91] "BU(R)"             "AL(L)"             "ROB(R)"           
#>  [94] "CRE(-ROB,-RUB)(R)" "CX"                "PB(L1)"           
#>  [97] "PLP(R)"            "EB"                "EPA(R)"           
#> [100] "GA(R)"             "PB(R3)"            "NO2(L)"           
#> [103] "dACA(R)"           "PB(R8)"            "PB(R9)"           
#> [106] "NO3(L)"            "a'3(R)"            "PED(R)"           
#> [109] "FLA(R)"            "LO(R)"             "CA(R)"            
#> [112] "AOT(R)"            "SMP(L)"            "a'L(R)"           
#> [115] "SAD(-AMMC)"        "LAL(R)"            "a1(R)"            
#> [118] "a'2(R)"            "LH(R)"             "AVLP(R)"          
#> [121] "a3(R)"             "IPS(R)"            "b'L(L)"           
#> [124] "VLNP(R)"           "AOTU(R)"           "CRE(L)"           
#> [127] "FBl9"              "PB(R4)"            "CA(L)"            
#> [130] "GOR(L)"            "EBr2r4"            "PB(L4)"           
#> [133] "SCL(L)"            "FBl7"              "OL(R)"            
#> [136] "PB(R5)"            "LAL(L)"            "GC"               
#> [139] "g2(R)"             "bL(L)"             "SNP(L)"           
#> [142] "PB(R2)"            "aL(L)"             "a'L(L)"           
#> [145] "g3(R)"             "CRE(R)"            "NO1(R)"           
#> [148] "gL(R)"             "SCL(R)"            "IB"               
#> [151] "AMMC"             
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "PB(R6)"            "dACA(R)"           "AL-VC5(R)"        
#>   [4] "AL-V(R)"           "CRE(R)"            "PB(L8)"           
#>   [7] "NO"                "CRE(-RUB)(L)"      "AL-DC2(R)"        
#>  [10] "AL-DM6(R)"         "g2(R)"             "SIP(R)"           
#>  [13] "GA(R)"             "NO1(R)"            "SCL(L)"           
#>  [16] "AL-DA2(L)"         "bL(L)"             "PVLP(R)"          
#>  [19] "g3(R)"             "AL-DA3(R)"         "AL-DC2(L)"        
#>  [22] "AL-DM2(L)"         "gL(L)"             "AL-VM7v(L)"       
#>  [25] "AL-DL2v(R)"        "CA(R)"             "mALT(R)"          
#>  [28] "AL-DA1(R)"         "AL-VP2(R)"         "PB"               
#>  [31] "AL-DL1(R)"         "AL-DM3(R)"         "a'3(R)"           
#>  [34] "PED(R)"            "NO1(L)"            "CAN(R)"           
#>  [37] "MB(R)"             "AL-VA1d(R)"        "FBl4"             
#>  [40] "PB(R8)"            "LH(R)"             "AL-VP3(R)"        
#>  [43] "vACA(R)"           "PENP"              "MB(L)"            
#>  [46] "NO3(R)"            "VMNP"              "PB(L6)"           
#>  [49] "a3(R)"             "AL-DA4l(R)"        "SPS(R)"           
#>  [52] "SPS(L)"            "FBl3"              "AL-VA7m(R)"       
#>  [55] "EBr3am"            "PB(L5)"            "LAL(-GA)(R)"      
#>  [58] "aL(R)"             "SNP(R)"            "EPA(L)"           
#>  [61] "AL-VA6(R)"         "LX(R)"             "a'1(R)"           
#>  [64] "FBl6"              "AL-DC3(R)"         "AL-DP1l(R)"       
#>  [67] "AOT(R)"            "GF(R)"             "SAD(-AMMC)"       
#>  [70] "AL-DM1(R)"         "AL-VC3l(R)"        "EBr6"             
#>  [73] "a'2(R)"            "NO(R)"             "AL-VL2a(R)"       
#>  [76] "b'1(R)"            "ATL(L)"            "SCL(R)"           
#>  [79] "AL-DP1m(L)"        "aL(L)"             "PB(R5)"           
#>  [82] "SMP(L)"            "PB(R7)"            "AL-DM5(L)"        
#>  [85] "GOR(R)"            "ATL(R)"            "bL(R)"            
#>  [88] "OL(R)"             "AL-VM5d(R)"        "PLP(R)"           
#>  [91] "AL-DM4(R)"         "WED(R)"            "EB"               
#>  [94] "AL-DL5(R)"         "AL-VM1(R)"         "AB(L)"            
#>  [97] "SAD"               "FLA(R)"            "NO3(L)"           
#> [100] "BU(L)"             "AL-VA2(R)"         "AL-VA5(R)"        
#> [103] "EBr3pw"            "PB(R3)"            "SLP(R)"           
#> [106] "AL-VM7v(R)"        "LOP(R)"            "AMMC"             
#> [109] "AL-DL5(L)"         "AL-VA6(L)"         "AL-VA1v(R)"       
#> [112] "b1(R)"             "b'L(L)"            "AL-DA4m(L)"       
#> [115] "IPS(R)"            "AL-D(R)"           "AL-DM1(L)"        
#> [118] "PB(L9)"            "AVLP(R)"           "a'L(L)"           
#> [121] "g1(R)"             "AL-D(L)"           "AL-DA2(R)"        
#> [124] "AL-DL4(R)"         "LX(L)"             "AL-DL3(R)"        
#> [127] "AL-DP1m(R)"        "b'2(R)"            "ICL(L)"           
#> [130] "NO2(R)"            "a1(R)"             "AL-VP1d(R)"       
#> [133] "EBr1"              "FB-column3"        "FBl1"             
#> [136] "FBl2"              "FBl5"              "AL-VM2(R)"        
#> [139] "GC"                "PB(L1)"            "AL-VA7l(R)"       
#> [142] "IB"                "b2(R)"             "AL-VL2p(R)"       
#> [145] "SNP(L)"            "ICL(R)"            "CRE(-ROB,-RUB)(R)"
#> [148] "RUB(L)"            "PB(R4)"            "NO(L)"            
#> [151] "PB(L4)"            "AL-VC3m(R)"        "PB(R1)"           
#> [154] "AME(R)"            "AL-DC4(R)"         "NO2(L)"           
#> [157] "SIP(L)"            "GOR(L)"            "g5(R)"            
#> [160] "INP"               "AL-DM5(R)"         "LAL(L)"           
#> [163] "POC"               "SMP(R)"            "hemibrain"        
#> [166] "ME(R)"             "AL-DA4m(R)"        "AL-VM7d(R)"       
#> [169] "lACA(R)"           "a'L(R)"            "FBl9"             
#> [172] "AL-VC1(R)"         "AL-VP1m(R)"        "VLNP(R)"          
#> [175] "MB(+ACA)(R)"       "AL-VA3(R)"         "PRW"              
#> [178] "AL-DM3(L)"         "PB(L7)"            "EPA(R)"           
#> [181] "GNG"               "AL-VC4(R)"         "AL-VP4(R)"        
#> [184] "b'L(R)"            "BU(R)"             "AL-VA4(R)"        
#> [187] "AL-VM7d(L)"        "PB(R9)"            "AL-DM2(R)"        
#> [190] "FB"                "AL(L)"             "AL-DL4(L)"        
#> [193] "FBl7"              "AB(R)"             "AL-VM5v(R)"       
#> [196] "EBr3d"             "AOTU(R)"           "ROB(R)"           
#> [199] "LAL(R)"            "AL-DC4(L)"         "g4(R)"            
#> [202] "gL(R)"             "VES(L)"            "AL-DL2d(R)"       
#> [205] "AL-VM3(R)"         "AL-VP1l(R)"        "mALT(L)"          
#> [208] "EBr2r4"            "AL-DA3(L)"         "PB(R2)"           
#> [211] "AL-DC1(R)"         "AL-VM4(R)"         "AL-VP5(R)"        
#> [214] "AL-VL1(R)"         "PB(L2)"            "a2(R)"            
#> [217] "EBr5"              "AL-DM4(L)"         "VES(R)"           
#> [220] "AL-VC2(R)"         "CX"                "CA(L)"            
#> [223] "PB(L3)"            "AL-DM6(L)"         "LO(R)"            
#> [226] "FBl8"              "AL(R)"             "RUB(R)"           
#> [229] "AL-DC1(L)"         "CRE(L)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AOTU(R)"           "PRW"               "NO1(L)"           
#>   [4] "mALT(R)"           "AL-VA6(L)"         "a'1(R)"           
#>   [7] "FBl9"              "CAN(R)"            "INP"              
#>  [10] "AL-VP1l(R)"        "PB(L3)"            "bL(R)"            
#>  [13] "AL-VP1m(R)"        "EBr3am"            "AL-DC2(L)"        
#>  [16] "VES(L)"            "AL-DL2d(R)"        "AL-VP2(R)"        
#>  [19] "AL-DM4(L)"         "CRE(L)"            "b'L(L)"           
#>  [22] "AL-VM5d(R)"        "SPS(R)"            "AL-DM2(L)"        
#>  [25] "AL-DC3(R)"         "ATL(R)"            "OL(R)"            
#>  [28] "AL(R)"             "AL-DL2v(R)"        "PB(R8)"           
#>  [31] "g4(R)"             "FBl6"              "AL-DL4(R)"        
#>  [34] "AL-DM6(R)"         "AOT(R)"            "PB(L6)"           
#>  [37] "AL-VL2a(R)"        "SMP(R)"            "SNP(L)"           
#>  [40] "b2(R)"             "FBl7"              "LH(R)"            
#>  [43] "SLP(R)"            "AL-VA1v(R)"        "AL-VC2(R)"        
#>  [46] "BU(R)"             "FBl5"              "LAL(R)"           
#>  [49] "AL-VA7l(R)"        "AL-VM2(R)"         "b'L(R)"           
#>  [52] "PB(L9)"            "VES(R)"            "AME(R)"           
#>  [55] "a1(R)"             "SAD(-AMMC)"        "AVLP(R)"          
#>  [58] "PB(R4)"            "AL-DC1(R)"         "PB(L2)"           
#>  [61] "NO1(R)"            "PB"                "NO2(R)"           
#>  [64] "AL-VA6(R)"         "PB(R6)"            "NO3(R)"           
#>  [67] "gL(R)"             "vACA(R)"           "FLA(R)"           
#>  [70] "NO(R)"             "NO3(L)"            "EBr3d"            
#>  [73] "SNP(R)"            "AL(L)"             "AL-DA4m(R)"       
#>  [76] "AL-DL5(R)"         "NO2(L)"            "LAL(-GA)(R)"      
#>  [79] "WED(R)"            "SCL(R)"            "FBl2"             
#>  [82] "AB(L)"             "GF(R)"             "AL-DL4(L)"        
#>  [85] "AL-DM3(L)"         "LX(L)"             "AL-DM3(R)"        
#>  [88] "AL-DM6(L)"         "IPS(R)"            "MB(R)"            
#>  [91] "AL-DM1(R)"         "BU(L)"             "EBr3pw"           
#>  [94] "AL-DA4l(R)"        "AL-VC5(R)"         "EBr2r4"           
#>  [97] "AL-DM4(R)"         "AL-VP5(R)"         "PB(L5)"           
#> [100] "a2(R)"             "g2(R)"             "AL-VP1d(R)"       
#> [103] "lACA(R)"           "PB(R7)"            "GC"               
#> [106] "EPA(R)"            "PB(R9)"            "LO(R)"            
#> [109] "CRE(-ROB,-RUB)(R)" "g3(R)"             "aL(R)"            
#> [112] "AL-D(L)"           "CRE(-RUB)(L)"      "FBl1"             
#> [115] "GNG"               "a'2(R)"            "PB(L7)"           
#> [118] "AB(R)"             "AL-DM5(R)"         "AL-DA2(L)"        
#> [121] "AL-VA1d(R)"        "PB(L8)"            "POC"              
#> [124] "CRE(R)"            "CA(R)"             "LOP(R)"           
#> [127] "SMP(L)"            "gL(L)"             "AL-VM7d(L)"       
#> [130] "AL-DL3(R)"         "SIP(L)"            "VMNP"             
#> [133] "AL-D(R)"           "AMMC"              "LAL(L)"           
#> [136] "hemibrain"         "SPS(L)"            "a'3(R)"           
#> [139] "AL-DA4m(L)"        "PB(R1)"            "AL-VC3m(R)"       
#> [142] "PVLP(R)"           "VLNP(R)"           "RUB(R)"           
#> [145] "SIP(R)"            "a3(R)"             "AL-DM2(R)"        
#> [148] "AL-VM7d(R)"        "PENP"              "ME(R)"            
#> [151] "SCL(L)"            "PLP(R)"            "EB"               
#> [154] "aL(L)"             "SAD"               "AL-DA1(R)"        
#> [157] "AL-VC3l(R)"        "PB(L4)"            "FBl4"             
#> [160] "AL-V(R)"           "MB(L)"             "AL-DA3(L)"        
#> [163] "AL-VP4(R)"         "b'1(R)"            "PED(R)"           
#> [166] "EBr1"              "ICL(R)"            "AL-DA3(R)"        
#> [169] "AL-VM3(R)"         "mALT(L)"           "GA(R)"            
#> [172] "AL-DM5(L)"         "ICL(L)"            "FB-column3"       
#> [175] "AL-DC4(R)"         "AL-VM4(R)"         "AL-VM7v(R)"       
#> [178] "AL-DC4(L)"         "EPA(L)"            "AL-VA2(R)"        
#> [181] "AL-VA7m(R)"        "MB(+ACA)(R)"       "FBl3"             
#> [184] "PB(R3)"            "AL-VA5(R)"         "GOR(L)"           
#> [187] "dACA(R)"           "IB"                "EBr5"             
#> [190] "PB(R2)"            "AL-VA3(R)"         "CX"               
#> [193] "FB"                "PB(L1)"            "AL-DC1(L)"        
#> [196] "AL-VC4(R)"         "b'2(R)"            "EBr6"             
#> [199] "bL(L)"             "b1(R)"             "NO(L)"            
#> [202] "AL-DL1(R)"         "AL-DP1m(L)"        "AL-VP3(R)"        
#> [205] "AL-DP1l(R)"        "AL-DL5(L)"         "AL-DC2(R)"        
#> [208] "ATL(L)"            "AL-VM7v(L)"        "GOR(R)"           
#> [211] "a'L(L)"            "AL-DP1m(R)"        "AL-VM1(R)"        
#> [214] "CA(L)"             "g1(R)"             "a'L(R)"           
#> [217] "AL-VA4(R)"         "AL-VL1(R)"         "AL-VL2p(R)"       
#> [220] "PB(R5)"            "LX(R)"             "AL-DM1(L)"        
#> [223] "ROB(R)"            "g5(R)"             "RUB(L)"           
#> [226] "AL-DA2(R)"         "FBl8"              "AL-VC1(R)"        
#> [229] "NO"                "AL-VM5v(R)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                          
#> [1,] "POST" "/api/v:ver/npexplorer/distribution"          
#> [2,] "POST" "/api/v:ver/raw/cypher/transaction/:id/commit"
#> [3,] "GET"  "/api/dbmeta/version"                         
#> [4,] "GET"  "/api/dbmeta/datasets"                        
#> [5,] "GET"  "/api/dbmeta/instances"                       
#> [6,] "GET"  "/api/v:ver/npexplorer/simpleconnections"     
#>      name                                                                                 
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                 
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                 
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getVersion-fm"      
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatasets-fm"     
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDataInstances-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                 
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }