Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "EBr2r4"            "AL-DM6(L)"         "NO(R)"            
#>   [4] "AL-DM5(R)"         "AL-VL1(R)"         "PB(L3)"           
#>   [7] "aL(L)"             "AL-VA1v(R)"        "NO2(L)"           
#>  [10] "NO1(R)"            "AL-VA4(R)"         "AL-VA6(R)"        
#>  [13] "PB(R5)"            "FBl4"              "INP"              
#>  [16] "PB(R4)"            "AL-VP2(R)"         "CA(L)"            
#>  [19] "AL-VA6(L)"         "mALT(R)"           "GC"               
#>  [22] "PB(R2)"            "lACA(R)"           "AL-VA2(R)"        
#>  [25] "b2(R)"             "SMP(L)"            "AL-DL2v(R)"       
#>  [28] "AL-DP1m(L)"        "PB(R9)"            "FB-column3"       
#>  [31] "b1(R)"             "AL-VP1m(R)"        "SCL(L)"           
#>  [34] "NO"                "AB(R)"             "GA(R)"            
#>  [37] "LX(L)"             "AL-VC2(R)"         "AL-VC5(R)"        
#>  [40] "FBl2"              "AL-VP4(R)"         "AL-DL5(R)"        
#>  [43] "NO2(R)"            "AL-VM1(R)"         "AL-VM2(R)"        
#>  [46] "LH(R)"             "PB(R8)"            "SIP(L)"           
#>  [49] "AL-DL4(L)"         "IB112"             "b'L(L)"           
#>  [52] "GF(R)"             "ICL(L)"            "AME(R)"           
#>  [55] "MB(R)"             "a'L(R)"            "FBl1"             
#>  [58] "VMNP"              "AL-DM2(L)"         "AL-VM7d(R)"       
#>  [61] "BU(L)"             "gL(L)"             "GOR(L)"           
#>  [64] "g4(R)"             "AL-DM4(L)"         "a'1(R)"           
#>  [67] "AL-VC3m(R)"        "AL-VP3(R)"         "hemibrain"        
#>  [70] "MB(+ACA)(R)"       "NO1(L)"            "AL-DA4m(L)"       
#>  [73] "FBl6"              "a3(R)"             "AL-DA1(R)"        
#>  [76] "AL-DC3(R)"         "AL-DL1(R)"         "AL-DL3(R)"        
#>  [79] "NO3(R)"            "AL-DM3(R)"         "AL-DP1l(R)"       
#>  [82] "CRE(R)"            "AL-DA4m(R)"        "AL(L)"            
#>  [85] "PB(L8)"            "PB(L6)"            "IPS(R)"           
#>  [88] "AL-VA7l(R)"        "AL-VM4(R)"         "b'L(R)"           
#>  [91] "AL-DL5(L)"         "PB(R6)"            "mALT(L)"          
#>  [94] "EPA(R)"            "EBr3pw"            "LX(R)"            
#>  [97] "CA(R)"             "AL-DL2d(R)"        "AL-D(R)"          
#> [100] "AL-VL2p(R)"        "FBl8"              "LAL(R)"           
#> [103] "AL-DA4l(R)"        "PB(R7)"            "VES(L)"           
#> [106] "AL-VM7v(R)"        "SNP(L)"            "SCL(R)"           
#> [109] "BU(R)"             "PB(R1)"            "AL-DP1m(R)"       
#> [112] "AL-VM5d(R)"        "SAD"               "AL-DM3(L)"        
#> [115] "SIP(R)"            "PB(L9)"            "AL-DL4(R)"        
#> [118] "AL-VC4(R)"         "PLP(R)"            "RUB(R)"           
#> [121] "AL-DM4(R)"         "AL-DC1(L)"         "AL-VM7d(L)"       
#> [124] "AL-VL2a(R)"        "AB(L)"             "PENP"             
#> [127] "a2(R)"             "PB(L7)"            "a'L(L)"           
#> [130] "AL-DC4(R)"         "AL-VM3(R)"         "AOTU(R)"          
#> [133] "CRE(L)"            "AL-DC2(R)"         "AL-DM2(R)"        
#> [136] "AL-VA7m(R)"        "AL-DA2(L)"         "AL-VM5v(R)"       
#> [139] "AL-DA2(R)"         "ATL(R)"            "bL(R)"            
#> [142] "PB"                "AL-DM1(R)"         "AL-VC1(R)"        
#> [145] "AL-VP1l(R)"        "AOT(R)"            "SMP(R)"           
#> [148] "vACA(R)"           "GOR(R)"            "LO(R)"            
#> [151] "ICL(R)"            "POC"               "g5(R)"            
#> [154] "CAN(R)"            "b'1(R)"            "PB(L2)"           
#> [157] "CX"                "LAL(L)"            "FLA(R)"           
#> [160] "AL-DM5(L)"         "bL(L)"             "PB(L4)"           
#> [163] "OL(R)"             "AL(R)"             "EBr1"             
#> [166] "gL(R)"             "VES(R)"            "AL-V(R)"          
#> [169] "AL-VP5(R)"         "IB"                "a'2(R)"           
#> [172] "FBl9"              "AL-DA3(R)"         "EBr3am"           
#> [175] "AL-DM1(L)"         "LOP(R)"            "AL-DM6(R)"        
#> [178] "AL-VA3(R)"         "ROB(R)"            "NO(L)"            
#> [181] "CRE(-ROB,-RUB)(R)" "a1(R)"             "PED(R)"           
#> [184] "EBr6"              "PRW"               "b'2(R)"           
#> [187] "dACA(R)"           "AL-DC4(L)"         "FBl7"             
#> [190] "AL-VA5(R)"         "g3(R)"             "AL-DC2(L)"        
#> [193] "FBl3"              "RUB(L)"            "GNG"              
#> [196] "AL-VC3l(R)"        "LAL(-GA)(R)"       "AL-VM7v(L)"       
#> [199] "AVLP(R)"           "SLP(R)"            "AL-DC1(R)"        
#> [202] "a'3(R)"            "ME(R)"             "EBr3d"            
#> [205] "SAD(-AMMC)"        "SPS(R)"            "PB(L5)"           
#> [208] "SPS(L)"            "AL-DA3(L)"         "AL-VP1d(R)"       
#> [211] "g1(R)"             "g2(R)"             "EBr5"             
#> [214] "SNP(R)"            "NO3(L)"            "FB"               
#> [217] "EB"                "AL-D(L)"           "CRE(-RUB)(L)"     
#> [220] "PVLP(R)"           "PB(R3)"            "AL-VA1d(R)"       
#> [223] "PB(L1)"            "MB(L)"             "AMMC"             
#> [226] "WED(R)"            "ATL(L)"            "VLNP(R)"          
#> [229] "FBl5"              "EPA(L)"            "aL(R)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "NO2(R)"            "EBr2r4"            "EBr3pw"           
#>   [4] "CA(R)"             "PB(R2)"            "a2(R)"            
#>   [7] "bL(R)"             "LAL(R)"            "POC"              
#>  [10] "CRE(-ROB,-RUB)(R)" "SIP(R)"            "PB(R1)"           
#>  [13] "FBl7"              "SLP(R)"            "GF(R)"            
#>  [16] "mALT(R)"           "a'L(R)"            "gL(L)"            
#>  [19] "FBl2"              "INP"               "VES(L)"           
#>  [22] "g3(R)"             "EBr1"              "SNP(R)"           
#>  [25] "PB(L7)"            "PB(R6)"            "FBl3"             
#>  [28] "PB(R5)"            "a'L(L)"            "WED(R)"           
#>  [31] "PB(L6)"            "FBl9"              "ICL(R)"           
#>  [34] "EBr3am"            "NO3(L)"            "PVLP(R)"          
#>  [37] "PB(R4)"            "PB(L2)"            "EPA(R)"           
#>  [40] "b1(R)"             "MB(R)"             "b'L(R)"           
#>  [43] "PB(R8)"            "a'1(R)"            "AVLP(R)"          
#>  [46] "CX"                "PRW"               "AL(L)"            
#>  [49] "VES(R)"            "VLNP(R)"           "a1(R)"            
#>  [52] "MB(L)"             "NO1(R)"            "EB"               
#>  [55] "PB(R9)"            "a'2(R)"            "EPA(L)"           
#>  [58] "PENP"              "SAD"               "CRE(R)"           
#>  [61] "PLP(R)"            "GOR(R)"            "SCL(L)"           
#>  [64] "NO1(L)"            "dACA(R)"           "LH(R)"            
#>  [67] "ATL(R)"            "EBr6"              "BU(L)"            
#>  [70] "bL(L)"             "VMNP"              "AME(R)"           
#>  [73] "RUB(L)"            "SMP(R)"            "GA(R)"            
#>  [76] "g5(R)"             "mALT(L)"           "CA(L)"            
#>  [79] "CRE(L)"            "PB(L5)"            "LAL(L)"           
#>  [82] "NO3(R)"            "PB(L1)"            "PB(L9)"           
#>  [85] "FB"                "a3(R)"             "NO(L)"            
#>  [88] "AB(L)"             "g4(R)"             "NO(R)"            
#>  [91] "SNP(L)"            "LOP(R)"            "PB(R3)"           
#>  [94] "GNG"               "a'3(R)"            "AOTU(R)"          
#>  [97] "SPS(L)"            "ME(R)"             "FBl6"             
#> [100] "ICL(L)"            "OL(R)"             "AOT(R)"           
#> [103] "IB"                "lACA(R)"           "LX(R)"            
#> [106] "BU(R)"             "NO"                "NO2(L)"           
#> [109] "PED(R)"            "FLA(R)"            "ROB(R)"           
#> [112] "gL(R)"             "SMP(L)"            "SCL(R)"           
#> [115] "LX(L)"             "PB(L4)"            "ATL(L)"           
#> [118] "AMMC"              "PB(R7)"            "GOR(L)"           
#> [121] "FBl1"              "CAN(R)"            "GC"               
#> [124] "LO(R)"             "vACA(R)"           "hemibrain"        
#> [127] "FBl5"              "AB(R)"             "IPS(R)"           
#> [130] "PB(L3)"            "g2(R)"             "PB(L8)"           
#> [133] "MB(+ACA)(R)"       "FBl8"              "b'2(R)"           
#> [136] "SIP(L)"            "aL(R)"             "b2(R)"            
#> [139] "b'L(L)"            "RUB(R)"            "b'1(R)"           
#> [142] "EBr5"              "PB"                "SAD(-AMMC)"       
#> [145] "FBl4"              "g1(R)"             "AL(R)"            
#> [148] "SPS(R)"            "LAL(-GA)(R)"       "EBr3d"            
#> [151] "aL(L)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DA4l(R)"        "AL-VA1d(R)"        "AL-VM7d(R)"       
#>   [4] "dACA(R)"           "PVLP(R)"           "OL(R)"            
#>   [7] "VES(L)"            "MB(R)"             "CA(L)"            
#>  [10] "PENP"              "FB"                "PB(R8)"           
#>  [13] "AL-DM1(L)"         "AL-DA4m(L)"        "PB(R9)"           
#>  [16] "MB(+ACA)(R)"       "AL-VM5v(R)"        "AL-VM7d(L)"       
#>  [19] "PB(L6)"            "AL-VP4(R)"         "PRW"              
#>  [22] "AL-VC1(R)"         "AL-VC3l(R)"        "AL-VC4(R)"        
#>  [25] "AL-VL1(R)"         "NO3(L)"            "PB(R4)"           
#>  [28] "AL-DM3(R)"         "AL-VC2(R)"         "a'L(R)"           
#>  [31] "AL-DM5(L)"         "AL-VM7v(L)"        "AL-DC4(R)"        
#>  [34] "PLP(R)"            "AL-DL5(R)"         "PED(R)"           
#>  [37] "SLP(R)"            "AL-DC3(R)"         "vACA(R)"          
#>  [40] "SMP(L)"            "AL-DM6(L)"         "AL-VM5d(R)"       
#>  [43] "VLNP(R)"           "RUB(R)"            "AOTU(R)"          
#>  [46] "g2(R)"             "GF(R)"             "AL-DL1(R)"        
#>  [49] "AL-DA1(R)"         "ICL(R)"            "AB(R)"            
#>  [52] "LX(L)"             "bL(R)"             "a'L(L)"           
#>  [55] "SNP(L)"            "IPS(R)"            "AL-VC5(R)"        
#>  [58] "BU(R)"             "AL-DP1m(L)"        "gL(R)"            
#>  [61] "CRE(L)"            "PB(L7)"            "EBr2r4"           
#>  [64] "LH(R)"             "GNG"               "AMMC"             
#>  [67] "SIP(R)"            "AL-VA6(R)"         "AL-VC3m(R)"       
#>  [70] "a'3(R)"            "SPS(R)"            "AL-DM3(L)"        
#>  [73] "CA(R)"             "ATL(R)"            "AL-VA3(R)"        
#>  [76] "AL-DC1(L)"         "EBr3pw"            "LOP(R)"           
#>  [79] "EBr6"              "FLA(R)"            "AME(R)"           
#>  [82] "AL-DP1m(R)"        "IB"                "AL-D(L)"          
#>  [85] "AL-DC2(R)"         "b'1(R)"            "AL-DA4m(R)"       
#>  [88] "AL-VA7m(R)"        "POC"               "AL-VA7l(R)"       
#>  [91] "AL-VM7v(R)"        "SMP(R)"            "bL(L)"            
#>  [94] "mALT(L)"           "PB(R7)"            "EBr5"             
#>  [97] "AL-DA3(L)"         "LO(R)"             "PB(R1)"           
#> [100] "b'L(L)"            "AL-VA4(R)"         "AL-DA2(L)"        
#> [103] "a3(R)"             "AL-VA1v(R)"        "b'2(R)"           
#> [106] "g3(R)"             "GC"                "AL-VA2(R)"        
#> [109] "AL-VM4(R)"         "ATL(L)"            "AL-VM2(R)"        
#> [112] "FBl5"              "CAN(R)"            "AL-DL2d(R)"       
#> [115] "hemibrain"         "lACA(R)"           "AL-DA2(R)"        
#> [118] "AB(L)"             "LAL(R)"            "PB"               
#> [121] "SNP(R)"            "AL-DM4(L)"         "CRE(-RUB)(L)"     
#> [124] "aL(L)"             "b2(R)"             "b'L(R)"           
#> [127] "ICL(L)"            "AL-DL3(R)"         "AL-VA5(R)"        
#> [130] "g1(R)"             "PB(R3)"            "AL-DA3(R)"        
#> [133] "AL-VL2p(R)"        "AL-VP1m(R)"        "AL-DL2v(R)"       
#> [136] "SAD"               "SCL(R)"            "PB(L8)"           
#> [139] "FBl6"              "FBl7"              "AL-D(R)"          
#> [142] "EB"                "SAD(-AMMC)"        "FBl9"             
#> [145] "INP"               "AL-DM2(R)"         "b1(R)"            
#> [148] "NO1(R)"            "LX(R)"             "g4(R)"            
#> [151] "EPA(L)"            "a1(R)"             "AL-VP3(R)"        
#> [154] "gL(L)"             "AL-VA6(L)"         "NO(R)"            
#> [157] "NO2(R)"            "RUB(L)"            "AL-VM3(R)"        
#> [160] "AL-DC2(L)"         "NO"                "FBl3"             
#> [163] "FBl8"              "GOR(R)"            "AL-DL4(R)"        
#> [166] "NO1(L)"            "FBl2"              "NO2(L)"           
#> [169] "AL-VP2(R)"         "AL-VP1d(R)"        "PB(L3)"           
#> [172] "WED(R)"            "a'1(R)"            "NO(L)"            
#> [175] "AL-DM4(R)"         "LAL(L)"            "g5(R)"            
#> [178] "PB(L4)"            "PB(L5)"            "CX"               
#> [181] "NO3(R)"            "AL-DC1(R)"         "AL-VP1l(R)"       
#> [184] "EBr3am"            "AL-VP5(R)"         "SPS(L)"           
#> [187] "ROB(R)"            "PB(R5)"            "VES(R)"           
#> [190] "AL-DM1(R)"         "SIP(L)"            "PB(L9)"           
#> [193] "FBl1"              "ME(R)"             "AL-DC4(L)"        
#> [196] "AL(R)"             "CRE(-ROB,-RUB)(R)" "mALT(R)"          
#> [199] "EBr3d"             "FB-column3"        "AOT(R)"           
#> [202] "SCL(L)"            "GA(R)"             "AL-VM1(R)"        
#> [205] "MB(L)"             "AVLP(R)"           "AL-DM5(R)"        
#> [208] "aL(R)"             "BU(L)"             "VMNP"             
#> [211] "AL-DL4(L)"         "PB(R2)"            "a'2(R)"           
#> [214] "AL-DL5(L)"         "AL-DM2(L)"         "FBl4"             
#> [217] "GOR(L)"            "EPA(R)"            "AL-V(R)"          
#> [220] "PB(L2)"            "CRE(R)"            "EBr1"             
#> [223] "AL(L)"             "AL-DM6(R)"         "a2(R)"            
#> [226] "PB(R6)"            "LAL(-GA)(R)"       "PB(L1)"           
#> [229] "AL-DP1l(R)"        "AL-VL2a(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-VM3(R)"         "AL-VA6(L)"         "PB(L8)"           
#>   [4] "ICL(L)"            "AL-VM7v(R)"        "NO1(L)"           
#>   [7] "g5(R)"             "AL-DC4(R)"         "AL-VA1v(R)"       
#>  [10] "aL(R)"             "AL-VC2(R)"         "PB(L4)"           
#>  [13] "EBr2r4"            "AL-DM1(L)"         "AL-DM6(L)"        
#>  [16] "AL-V(R)"           "AB(L)"             "PB(L1)"           
#>  [19] "AL-DL1(R)"         "CA(L)"             "AL-DA1(R)"        
#>  [22] "EB"                "NO3(L)"            "AL(R)"            
#>  [25] "AL-VA4(R)"         "SIP(L)"            "a'L(R)"           
#>  [28] "AL-D(L)"           "a'1(R)"            "b2(R)"            
#>  [31] "ME(R)"             "SIP(R)"            "FBl1"             
#>  [34] "AL-VL2p(R)"        "PED(R)"            "NO1(R)"           
#>  [37] "g2(R)"             "AL-DM3(L)"         "SNP(L)"           
#>  [40] "MB(R)"             "AL-DM3(R)"         "PRW"              
#>  [43] "AL-VP3(R)"         "PB(R4)"            "FBl4"             
#>  [46] "AL-VL1(R)"         "NO3(R)"            "SAD(-AMMC)"       
#>  [49] "ATL(R)"            "AL-DM2(R)"         "AL-DP1m(R)"       
#>  [52] "AL-DC1(L)"         "LO(R)"             "AL-DL5(R)"        
#>  [55] "AL-VM2(R)"         "AL-DA4m(L)"        "CX"               
#>  [58] "EBr3am"            "PVLP(R)"           "FBl5"             
#>  [61] "LAL(R)"            "AL-VA7l(R)"        "b'1(R)"           
#>  [64] "g4(R)"             "MB(+ACA)(R)"       "AB(R)"            
#>  [67] "NO(L)"             "GA(R)"             "FB-column3"       
#>  [70] "LH(R)"             "AL-VA5(R)"         "SMP(R)"           
#>  [73] "GF(R)"             "PB(R2)"            "OL(R)"            
#>  [76] "PLP(R)"            "SCL(L)"            "AL-DA4m(R)"       
#>  [79] "a'2(R)"            "PB(R1)"            "VES(R)"           
#>  [82] "FB"                "AL-VL2a(R)"        "CRE(-ROB,-RUB)(R)"
#>  [85] "NO2(L)"            "FBl8"              "LX(L)"            
#>  [88] "VLNP(R)"           "MB(L)"             "AL-DP1l(R)"       
#>  [91] "AL-DA4l(R)"        "AL-VA2(R)"         "a1(R)"            
#>  [94] "AL-VM7d(L)"        "vACA(R)"           "CRE(L)"           
#>  [97] "AL-DA3(L)"         "GOR(R)"            "EBr3d"            
#> [100] "CAN(R)"            "bL(R)"             "AL-DL2v(R)"       
#> [103] "AL-VP5(R)"         "AL-VP1l(R)"        "lACA(R)"          
#> [106] "PB(L3)"            "FBl2"              "FBl7"             
#> [109] "SLP(R)"            "AL-DL4(R)"         "PB(L2)"           
#> [112] "AMMC"              "PENP"              "WED(R)"           
#> [115] "VMNP"              "gL(R)"             "SMP(L)"           
#> [118] "FLA(R)"            "dACA(R)"           "PB"               
#> [121] "PB(R3)"            "AME(R)"            "EPA(L)"           
#> [124] "AL-DP1m(L)"        "SPS(L)"            "AL(L)"            
#> [127] "AL-DL4(L)"         "AL-D(R)"           "CRE(R)"           
#> [130] "CA(R)"             "a'L(L)"            "a'3(R)"           
#> [133] "AL-DM5(L)"         "IPS(R)"            "AL-DA2(R)"        
#> [136] "LAL(L)"            "AL-DM4(L)"         "RUB(L)"           
#> [139] "AL-VA7m(R)"        "AL-VC3l(R)"        "g3(R)"            
#> [142] "FBl3"              "IB"                "PB(R7)"           
#> [145] "SCL(R)"            "EBr5"              "bL(L)"            
#> [148] "AOTU(R)"           "AL-DC2(L)"         "AL-DC4(L)"        
#> [151] "EBr3pw"            "NO2(R)"            "VES(L)"           
#> [154] "LOP(R)"            "PB(R6)"            "g1(R)"            
#> [157] "PB(L7)"            "AL-DL3(R)"         "PB(L5)"           
#> [160] "mALT(R)"           "AL-DL2d(R)"        "gL(L)"            
#> [163] "AL-DM6(R)"         "aL(L)"             "PB(L6)"           
#> [166] "NO(R)"             "PB(L9)"            "AL-VM4(R)"        
#> [169] "PB(R5)"            "BU(L)"             "EBr1"             
#> [172] "AVLP(R)"           "AL-DM1(R)"         "AOT(R)"           
#> [175] "SPS(R)"            "b'L(R)"            "ROB(R)"           
#> [178] "FBl9"              "AL-VM7d(R)"        "b'2(R)"           
#> [181] "PB(R9)"            "AL-VM1(R)"         "BU(R)"            
#> [184] "AL-VC4(R)"         "AL-VC5(R)"         "AL-DC3(R)"        
#> [187] "AL-DA3(R)"         "PB(R8)"            "LX(R)"            
#> [190] "POC"               "AL-DC2(R)"         "GNG"              
#> [193] "AL-VC3m(R)"        "AL-VM5d(R)"        "EPA(R)"           
#> [196] "AL-DL5(L)"         "AL-VA1d(R)"        "AL-VC1(R)"        
#> [199] "AL-VP2(R)"         "a2(R)"             "CRE(-RUB)(L)"     
#> [202] "INP"               "AL-DM5(R)"         "AL-VM7v(L)"       
#> [205] "AL-DA2(L)"         "SAD"               "SNP(R)"           
#> [208] "b1(R)"             "a3(R)"             "GC"               
#> [211] "AL-VP1m(R)"        "RUB(R)"            "AL-VA3(R)"        
#> [214] "NO"                "b'L(L)"            "AL-DM4(R)"        
#> [217] "AL-VM5v(R)"        "hemibrain"         "AL-DM2(L)"        
#> [220] "FBl6"              "AL-DC1(R)"         "AL-VA6(R)"        
#> [223] "AL-VP4(R)"         "ATL(L)"            "LAL(-GA)(R)"      
#> [226] "EBr6"              "mALT(L)"           "GOR(L)"           
#> [229] "AL-VP1d(R)"        "ICL(R)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                           
#> [1,] "GET"  "/api/v:ver/npexplorer/rankedtable"            
#> [2,] "POST" "/api/npexplorer/roisinneuron"                 
#> [3,] "POST" "/api/v:ver/npexplorer/roisinneuron"           
#> [4,] "GET"  "/api/v:ver/npexplorer/celltype/:dataset/:type"
#> [5,] "GET"  "/api/cached/roiconnectivity"                  
#> [6,] "HEAD" "/api/*"                                       
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROIConnectivity-fm"    
#> [6,] "github.com/labstack/echo.glob..func1"                                                      

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }