Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "lACA(R)"           "b'L(R)"            "GNG"              
#>   [4] "AL-VC1(R)"         "aL(L)"             "INP"              
#>   [7] "mALT(R)"           "EBr3d"             "AL-VA6(L)"        
#>  [10] "CRE(L)"            "SAD"               "b1(R)"            
#>  [13] "AL-DM1(R)"         "AL-DA4l(R)"        "AL-DL5(R)"        
#>  [16] "AL-DP1m(R)"        "SPS(R)"            "CRE(R)"           
#>  [19] "AL-D(L)"           "a'L(L)"            "FBl8"             
#>  [22] "AL-VA7m(R)"        "EBr2r4"            "ICL(L)"           
#>  [25] "LH(R)"             "AL-DM5(R)"         "AL-DL5(L)"        
#>  [28] "AL-DC1(R)"         "AL-DM6(R)"         "AL-VM7v(R)"       
#>  [31] "AL-VP1d(R)"        "SMP(L)"            "AL-DM2(L)"        
#>  [34] "FBl1"              "AL-VA3(R)"         "AL-VA5(R)"        
#>  [37] "AL-VM7v(L)"        "PB(L3)"            "VES(R)"           
#>  [40] "a1(R)"             "PB(R7)"            "CRE(-RUB)(L)"     
#>  [43] "FBl4"              "SCL(L)"            "AL-DM3(R)"        
#>  [46] "AL-VM3(R)"         "LAL(-GA)(R)"       "FLA(R)"           
#>  [49] "AL-DA3(L)"         "IPS(R)"            "AL-VA4(R)"        
#>  [52] "FBl3"              "AL-DP1m(L)"        "AL-DL2v(R)"       
#>  [55] "ROB(R)"            "IB112"             "AME(R)"           
#>  [58] "AL-VA6(R)"         "PB(R6)"            "aL(R)"            
#>  [61] "EBr6"              "AL-DM1(L)"         "PLP(R)"           
#>  [64] "GOR(R)"            "EBr3am"            "MB(R)"            
#>  [67] "PB(R5)"            "PB(L5)"            "PRW"              
#>  [70] "GC"                "SAD(-AMMC)"        "FBl2"             
#>  [73] "AL-DA1(R)"         "VLNP(R)"           "PB(L7)"           
#>  [76] "AB(R)"             "PB(R4)"            "LAL(R)"           
#>  [79] "CA(L)"             "PB(L8)"            "gL(L)"            
#>  [82] "AL-DM4(L)"         "AL-VC4(R)"         "LX(R)"            
#>  [85] "AL(L)"             "EBr3pw"            "PB(R3)"           
#>  [88] "gL(R)"             "SNP(L)"            "NO(L)"            
#>  [91] "CRE(-ROB,-RUB)(R)" "a'2(R)"            "AL-V(R)"          
#>  [94] "AOTU(R)"           "ME(R)"             "NO3(L)"           
#>  [97] "AL-VC3l(R)"        "AL-DA3(R)"         "AL-DA4m(R)"       
#> [100] "PB(L2)"            "MB(+ACA)(R)"       "AL-VA7l(R)"       
#> [103] "NO3(R)"            "LOP(R)"            "AL-VL1(R)"        
#> [106] "AL-VP1l(R)"        "RUB(R)"            "mALT(L)"          
#> [109] "EPA(R)"            "FBl5"              "AL-DL2d(R)"       
#> [112] "AL-VC3m(R)"        "AL-VA1v(R)"        "AL-VM1(R)"        
#> [115] "GF(R)"             "FB"                "g4(R)"            
#> [118] "NO(R)"             "GA(R)"             "AL-DA2(R)"        
#> [121] "AL-DL1(R)"         "IB"                "PB(L9)"           
#> [124] "b'L(L)"            "AL-DL3(R)"         "AL-VM5d(R)"       
#> [127] "dACA(R)"           "PB(R2)"            "CAN(R)"           
#> [130] "AL-VC5(R)"         "b'2(R)"            "SIP(L)"           
#> [133] "AL-DC1(L)"         "AL-VP3(R)"         "PB(L1)"           
#> [136] "ICL(R)"            "AL-VP1m(R)"        "AL-DC2(R)"        
#> [139] "AL-DM6(L)"         "ATL(R)"            "bL(R)"            
#> [142] "VES(L)"            "AL-DC2(L)"         "AL-VA1d(R)"       
#> [145] "ATL(L)"            "AL-DC4(L)"         "EPA(L)"           
#> [148] "SLP(R)"            "FBl9"              "a2(R)"            
#> [151] "EBr1"              "PB(R8)"            "PB(L6)"           
#> [154] "WED(R)"            "AL-VL2p(R)"        "hemibrain"        
#> [157] "b2(R)"             "AL-DL4(L)"         "LX(L)"            
#> [160] "SIP(R)"            "OL(R)"             "AL-DC3(R)"        
#> [163] "AL-DL4(R)"         "AL-DM4(R)"         "AB(L)"            
#> [166] "SCL(R)"            "VMNP"              "AL-DA2(L)"        
#> [169] "FBl7"              "NO1(L)"            "a'1(R)"           
#> [172] "bL(L)"             "NO2(R)"            "AL(R)"            
#> [175] "PB(R1)"            "g5(R)"             "AVLP(R)"          
#> [178] "AL-VM7d(R)"        "vACA(R)"           "NO1(R)"           
#> [181] "SPS(L)"            "NO"                "AL-DM2(R)"        
#> [184] "GOR(L)"            "PB"                "b'1(R)"           
#> [187] "LAL(L)"            "g3(R)"             "PVLP(R)"          
#> [190] "a'3(R)"            "g2(R)"             "CA(R)"            
#> [193] "AL-DP1l(R)"        "RUB(L)"            "AL-VM4(R)"        
#> [196] "POC"               "AOT(R)"            "AL-VA2(R)"        
#> [199] "EBr5"              "LO(R)"             "AL-VP4(R)"        
#> [202] "SMP(R)"            "SNP(R)"            "PB(R9)"           
#> [205] "AL-DC4(R)"         "AMMC"              "MB(L)"            
#> [208] "AL-VM7d(L)"        "AL-VL2a(R)"        "AL-VM2(R)"        
#> [211] "PENP"              "g1(R)"             "AL-D(R)"          
#> [214] "AL-VM5v(R)"        "BU(L)"             "PB(L4)"           
#> [217] "CX"                "NO2(L)"            "a'L(R)"           
#> [220] "BU(R)"             "AL-VP5(R)"         "a3(R)"            
#> [223] "AL-VC2(R)"         "AL-VP2(R)"         "AL-DA4m(L)"       
#> [226] "AL-DM5(L)"         "FB-column3"        "FBl6"             
#> [229] "AL-DM3(L)"         "PED(R)"            "EB"               
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "b'1(R)"            "SCL(L)"            "PB(L3)"           
#>   [4] "PED(R)"            "mALT(R)"           "NO1(R)"           
#>   [7] "a'L(R)"            "PB(L9)"            "SIP(L)"           
#>  [10] "IPS(R)"            "VLNP(R)"           "SAD"              
#>  [13] "PRW"               "ME(R)"             "NO2(L)"           
#>  [16] "GC"                "NO3(R)"            "SCL(R)"           
#>  [19] "ATL(R)"            "MB(R)"             "PB(L5)"           
#>  [22] "a3(R)"             "POC"               "b2(R)"            
#>  [25] "bL(L)"             "NO3(L)"            "ICL(L)"           
#>  [28] "LH(R)"             "LX(L)"             "PB(L2)"           
#>  [31] "LOP(R)"            "g3(R)"             "MB(L)"            
#>  [34] "gL(L)"             "BU(R)"             "SPS(L)"           
#>  [37] "b'L(L)"            "OL(R)"             "RUB(R)"           
#>  [40] "INP"               "GNG"               "lACA(R)"          
#>  [43] "WED(R)"            "dACA(R)"           "gL(R)"            
#>  [46] "FBl2"              "PB(R2)"            "SPS(R)"           
#>  [49] "PB(L1)"            "PB(R8)"            "NO(L)"            
#>  [52] "LAL(R)"            "a2(R)"             "FBl3"             
#>  [55] "ATL(L)"            "vACA(R)"           "CA(L)"            
#>  [58] "PB(L8)"            "a'2(R)"            "EBr3d"            
#>  [61] "AL(L)"             "CAN(R)"            "AB(L)"            
#>  [64] "CRE(R)"            "SNP(R)"            "SNP(L)"           
#>  [67] "FBl1"              "a'3(R)"            "b'2(R)"           
#>  [70] "hemibrain"         "SIP(R)"            "EBr3am"           
#>  [73] "AMMC"              "GF(R)"             "ICL(R)"           
#>  [76] "LAL(-GA)(R)"       "PB(L7)"            "AVLP(R)"          
#>  [79] "mALT(L)"           "EB"                "EPA(R)"           
#>  [82] "PB(R9)"            "FBl7"              "b'L(R)"           
#>  [85] "PVLP(R)"           "bL(R)"             "a'L(L)"           
#>  [88] "g1(R)"             "SLP(R)"            "AL(R)"            
#>  [91] "SMP(R)"            "PB(R1)"            "FBl4"             
#>  [94] "VES(L)"            "EBr5"              "PLP(R)"           
#>  [97] "EBr3pw"            "FBl5"              "AOT(R)"           
#> [100] "SMP(L)"            "g2(R)"             "LX(R)"            
#> [103] "AB(R)"             "GOR(L)"            "FBl9"             
#> [106] "AME(R)"            "PB(L4)"            "aL(R)"            
#> [109] "EBr6"              "a'1(R)"            "CX"               
#> [112] "PB(R6)"            "BU(L)"             "NO2(R)"           
#> [115] "FBl6"              "aL(L)"             "EBr2r4"           
#> [118] "NO(R)"             "RUB(L)"            "AOTU(R)"          
#> [121] "FB"                "PB"                "IB"               
#> [124] "VES(R)"            "FBl8"              "NO1(L)"           
#> [127] "EBr1"              "VMNP"              "CRE(L)"           
#> [130] "g5(R)"             "SAD(-AMMC)"        "PB(R5)"           
#> [133] "a1(R)"             "g4(R)"             "PB(R3)"           
#> [136] "CRE(-ROB,-RUB)(R)" "GA(R)"             "NO"               
#> [139] "PB(R4)"            "LAL(L)"            "PB(R7)"           
#> [142] "LO(R)"             "MB(+ACA)(R)"       "EPA(L)"           
#> [145] "GOR(R)"            "PENP"              "FLA(R)"           
#> [148] "ROB(R)"            "PB(L6)"            "CA(R)"            
#> [151] "b1(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "SAD(-AMMC)"        "AL-DA1(R)"         "PB(R7)"           
#>   [4] "EB"                "NO1(R)"            "AL-DL5(R)"        
#>   [7] "PB(L6)"            "ICL(R)"            "GNG"              
#>  [10] "g2(R)"             "ROB(R)"            "bL(L)"            
#>  [13] "FBl8"              "RUB(L)"            "AL-DA2(L)"        
#>  [16] "AL-DA3(L)"         "AL-VM1(R)"         "a1(R)"            
#>  [19] "GF(R)"             "AL-DL3(R)"         "AL-VC4(R)"        
#>  [22] "AL-DA2(R)"         "RUB(R)"            "CA(L)"            
#>  [25] "WED(R)"            "AL-DM1(L)"         "AOTU(R)"          
#>  [28] "b1(R)"             "AL-DP1l(R)"        "AL-VM5d(R)"       
#>  [31] "AMMC"              "NO2(L)"            "AVLP(R)"          
#>  [34] "aL(L)"             "NO3(L)"            "b'L(R)"           
#>  [37] "NO"                "g3(R)"             "LAL(R)"           
#>  [40] "AL-VP3(R)"         "a'1(R)"            "AL-DL2v(R)"       
#>  [43] "AL-DM3(R)"         "AB(L)"             "AL-DL5(L)"        
#>  [46] "AL-D(L)"           "AL-DC1(L)"         "AL-VM2(R)"        
#>  [49] "FLA(R)"            "g4(R)"             "AL-VM4(R)"        
#>  [52] "FBl6"              "AL-DM6(R)"         "AL-VP1m(R)"       
#>  [55] "AL-VL1(R)"         "LOP(R)"            "mALT(L)"          
#>  [58] "SLP(R)"            "AL-VM7d(R)"        "NO(L)"            
#>  [61] "GC"                "IPS(R)"            "AL-VA1d(R)"       
#>  [64] "AL-VA7l(R)"        "VES(R)"            "CAN(R)"           
#>  [67] "PB(R1)"            "EBr3pw"            "AL-DM4(L)"        
#>  [70] "AL-DL2d(R)"        "AOT(R)"            "GOR(L)"           
#>  [73] "NO3(R)"            "AL-DA4m(R)"        "AL-VP1d(R)"       
#>  [76] "VMNP"              "FBl9"              "PB(L7)"           
#>  [79] "PB(L8)"            "AL-DM5(L)"         "a'3(R)"           
#>  [82] "AL-VA4(R)"         "AL-VC3m(R)"        "GOR(R)"           
#>  [85] "AL-DC2(L)"         "LX(L)"             "FBl7"             
#>  [88] "AB(R)"             "AL-VA3(R)"         "AL-VA5(R)"        
#>  [91] "VLNP(R)"           "g1(R)"             "LX(R)"            
#>  [94] "FB-column3"        "CX"                "NO1(L)"           
#>  [97] "SIP(R)"            "EPA(L)"            "mALT(R)"          
#> [100] "LO(R)"             "AL-DL4(L)"         "CA(R)"            
#> [103] "NO2(R)"            "AL-VA6(R)"         "PB(L2)"           
#> [106] "vACA(R)"           "PED(R)"            "EBr5"             
#> [109] "AL-DC2(R)"         "AL-VA1v(R)"        "AL-VC3l(R)"       
#> [112] "EPA(R)"            "SNP(L)"            "AL-DL1(R)"        
#> [115] "CRE(L)"            "bL(R)"             "AME(R)"           
#> [118] "ME(R)"             "PB(L9)"            "AL(R)"            
#> [121] "AL-DM4(R)"         "AL-VM3(R)"         "AL-VP5(R)"        
#> [124] "EBr3am"            "EBr3d"             "AL-DM6(L)"        
#> [127] "PB(R5)"            "CRE(R)"            "AL-D(R)"          
#> [130] "EBr1"              "PB(L5)"            "LAL(L)"           
#> [133] "AL-VP4(R)"         "OL(R)"             "SMP(R)"           
#> [136] "IB"                "POC"               "b'1(R)"           
#> [139] "LAL(-GA)(R)"       "PB(R9)"            "MB(R)"            
#> [142] "FBl4"              "LH(R)"             "AL-VM7d(L)"       
#> [145] "PLP(R)"            "AL-VA6(L)"         "AL-VC5(R)"        
#> [148] "FB"                "PB(L4)"            "AL-VL2a(R)"       
#> [151] "a'2(R)"            "gL(R)"             "a'L(R)"           
#> [154] "AL(L)"             "PB(R3)"            "VES(L)"           
#> [157] "AL-VL2p(R)"        "ATL(L)"            "SIP(L)"           
#> [160] "EBr2r4"            "AL-DC4(R)"         "AL-DM2(R)"        
#> [163] "AL-V(R)"           "SCL(L)"            "PB(R2)"           
#> [166] "SAD"               "SNP(R)"            "AL-DP1m(L)"       
#> [169] "AL-VA7m(R)"        "AL-VM7v(R)"        "lACA(R)"          
#> [172] "AL-DA4l(R)"        "ICL(L)"            "AL-DC1(R)"        
#> [175] "AL-DP1m(R)"        "PENP"              "PB(L3)"           
#> [178] "AL-DA3(R)"         "SCL(R)"            "AL-DC4(L)"        
#> [181] "gL(L)"             "g5(R)"             "BU(R)"            
#> [184] "AL-DL4(R)"         "PRW"               "BU(L)"            
#> [187] "AL-DC3(R)"         "AL-VC1(R)"         "AL-VM5v(R)"       
#> [190] "aL(R)"             "SPS(L)"            "AL-DA4m(L)"       
#> [193] "a3(R)"             "AL-VC2(R)"         "SPS(R)"           
#> [196] "PB"                "AL-VM7v(L)"        "AL-VP1l(R)"       
#> [199] "a2(R)"             "AL-DM2(L)"         "AL-VA2(R)"        
#> [202] "AL-DM3(L)"         "PVLP(R)"           "FBl3"             
#> [205] "FBl5"              "INP"               "b'2(R)"           
#> [208] "b2(R)"             "PB(R8)"            "NO(R)"            
#> [211] "AL-DM1(R)"         "AL-DM5(R)"         "ATL(R)"           
#> [214] "PB(R4)"            "CRE(-RUB)(L)"      "FBl2"             
#> [217] "EBr6"              "PB(L1)"            "FBl1"             
#> [220] "b'L(L)"            "a'L(L)"            "AL-VP2(R)"        
#> [223] "hemibrain"         "MB(L)"             "GA(R)"            
#> [226] "dACA(R)"           "MB(+ACA)(R)"       "CRE(-ROB,-RUB)(R)"
#> [229] "PB(R6)"            "SMP(L)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "b'L(R)"            "AVLP(R)"           "AL-DA4l(R)"       
#>   [4] "SIP(R)"            "FB-column3"        "aL(L)"            
#>   [7] "NO(L)"             "SMP(L)"            "EB"               
#>  [10] "AL-DM2(L)"         "AL-DM5(L)"         "IB"               
#>  [13] "vACA(R)"           "GOR(L)"            "EPA(R)"           
#>  [16] "EBr3pw"            "AL-VM7d(R)"        "AL-VA5(R)"        
#>  [19] "AL-VM1(R)"         "AL-DA2(L)"         "FBl5"             
#>  [22] "PB(L4)"            "AL-VA1d(R)"        "PB(L9)"           
#>  [25] "ICL(L)"            "CAN(R)"            "BU(L)"            
#>  [28] "EBr2r4"            "AL-DM1(L)"         "a'1(R)"           
#>  [31] "CX"                "a'L(R)"            "FBl9"             
#>  [34] "LH(R)"             "AL-DM1(R)"         "AL-VP2(R)"        
#>  [37] "CRE(R)"            "AL-D(L)"           "PB(R9)"           
#>  [40] "AL-DL2v(R)"        "AL-DA4m(L)"        "AL-VC1(R)"        
#>  [43] "AMMC"              "AL-DC3(R)"         "CRE(-ROB,-RUB)(R)"
#>  [46] "LAL(-GA)(R)"       "AME(R)"            "EPA(L)"           
#>  [49] "VES(L)"            "a'2(R)"            "ATL(L)"           
#>  [52] "b1(R)"             "SPS(R)"            "aL(R)"            
#>  [55] "AB(R)"             "POC"               "a2(R)"            
#>  [58] "CRE(L)"            "SLP(R)"            "PB(L8)"           
#>  [61] "AL-VP3(R)"         "PB(L1)"            "FBl6"             
#>  [64] "AL-DA4m(R)"        "AL-DM3(R)"         "AL-VL1(R)"        
#>  [67] "AL-VP1m(R)"        "AL-DL4(L)"         "a'L(L)"           
#>  [70] "AL-DA1(R)"         "AL-VC3l(R)"        "SPS(L)"           
#>  [73] "AL-DM6(L)"         "PB(R4)"            "AL-VA1v(R)"       
#>  [76] "AL-VA6(R)"         "PB(R8)"            "PB(L6)"           
#>  [79] "GNG"               "INP"               "GF(R)"            
#>  [82] "PLP(R)"            "VMNP"              "SNP(L)"           
#>  [85] "PRW"               "MB(L)"             "g5(R)"            
#>  [88] "FBl4"              "FBl3"              "AL-VA3(R)"        
#>  [91] "AL-VP1l(R)"        "PB(R5)"            "WED(R)"           
#>  [94] "AL-VC3m(R)"        "SMP(R)"            "FLA(R)"           
#>  [97] "AL(R)"             "AL-VM3(R)"         "g4(R)"            
#> [100] "AL-DM3(L)"         "LO(R)"             "EBr6"             
#> [103] "AL-DL4(R)"         "AL-VL2a(R)"        "AL-DL2d(R)"       
#> [106] "AL-VC5(R)"         "PENP"              "EBr1"             
#> [109] "CRE(-RUB)(L)"      "FBl7"              "g1(R)"            
#> [112] "PB"                "AL-DM4(L)"         "EBr3am"           
#> [115] "AL-DA3(R)"         "AL-V(R)"           "a1(R)"            
#> [118] "EBr5"              "AL-DC4(L)"         "RUB(L)"           
#> [121] "MB(+ACA)(R)"       "FBl8"              "CA(R)"            
#> [124] "AL-VM7v(R)"        "b'1(R)"            "AL-DL5(L)"        
#> [127] "IPS(R)"            "AL-DL1(R)"         "AL-DM5(R)"        
#> [130] "AL-DM2(R)"         "a'3(R)"            "mALT(R)"          
#> [133] "NO(R)"             "PB(R3)"            "b'L(L)"           
#> [136] "SIP(L)"            "NO2(L)"            "PB(L3)"           
#> [139] "LOP(R)"            "PB(R1)"            "a3(R)"            
#> [142] "SCL(L)"            "AB(L)"             "FB"               
#> [145] "bL(R)"             "LAL(R)"            "AL-DP1l(R)"       
#> [148] "AL-DL5(R)"         "AL-VL2p(R)"        "PB(L5)"           
#> [151] "g3(R)"             "SAD"               "PB(L7)"           
#> [154] "OL(R)"             "VLNP(R)"           "CA(L)"            
#> [157] "b'2(R)"            "NO1(L)"            "AL-DA3(L)"        
#> [160] "AL-DC1(L)"         "hemibrain"         "AL-VM7v(L)"       
#> [163] "AL-DM6(R)"         "ICL(R)"            "GC"               
#> [166] "MB(R)"             "AL-VA4(R)"         "PB(R2)"           
#> [169] "AOTU(R)"           "AL-VM7d(L)"        "NO"               
#> [172] "NO3(L)"            "g2(R)"             "bL(L)"            
#> [175] "AOT(R)"            "FBl1"              "GOR(R)"           
#> [178] "AL-DA2(R)"         "AL-VA7l(R)"        "gL(R)"            
#> [181] "VES(R)"            "AL-VM5d(R)"        "AL-VP4(R)"        
#> [184] "lACA(R)"           "NO2(R)"            "LAL(L)"           
#> [187] "NO3(R)"            "EBr3d"             "SAD(-AMMC)"       
#> [190] "AL-VM5v(R)"        "PB(L2)"            "mALT(L)"          
#> [193] "PED(R)"            "gL(L)"             "AL(L)"            
#> [196] "AL-DL3(R)"         "NO1(R)"            "LX(L)"            
#> [199] "AL-DC2(R)"         "AL-DM4(R)"         "AL-VC2(R)"        
#> [202] "GA(R)"             "AL-DP1m(L)"        "ATL(R)"           
#> [205] "AL-VA2(R)"         "AL-VM4(R)"         "PB(R6)"           
#> [208] "AL-D(R)"           "AL-VM2(R)"         "AL-VA6(L)"        
#> [211] "FBl2"              "AL-DP1m(R)"        "SNP(R)"           
#> [214] "AL-DC1(R)"         "RUB(R)"            "PVLP(R)"          
#> [217] "AL-VC4(R)"         "b2(R)"             "dACA(R)"          
#> [220] "LX(R)"             "BU(R)"             "AL-VA7m(R)"       
#> [223] "AL-VP5(R)"         "AL-VP1d(R)"        "PB(R7)"           
#> [226] "SCL(R)"            "ME(R)"             "AL-DC4(R)"        
#> [229] "AL-DC2(L)"         "ROB(R)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method     path                               
#> [1,] "REPORT"   "/api"                             
#> [2,] "PROPFIND" "/api/*"                           
#> [3,] "GET"      "/api/v:ver/npexplorer/findneurons"
#> [4,] "GET"      "/api/npexplorer/neuronmeta"       
#> [5,] "DELETE"   "/api/*"                           
#> [6,] "TRACE"    "/api/*"                           
#>      name                                                                                      
#> [1,] "github.com/labstack/echo.glob..func1"                                                    
#> [2,] "github.com/labstack/echo.glob..func1"                                                    
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                      
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMeta-fm"
#> [5,] "github.com/labstack/echo.glob..func1"                                                    
#> [6,] "github.com/labstack/echo.glob..func1"                                                    

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }