Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PED(R)"            "ICL(L)"            "CA(R)"            
#>   [4] "SCL(L)"            "FLA(R)"            "SNP(R)"           
#>   [7] "SAD(-AMMC)"        "MB(+ACA)(R)"       "FBl7"             
#>  [10] "SMP(R)"            "PB(L5)"            "WED(R)"           
#>  [13] "NO3(L)"            "g5(R)"             "PB(R8)"           
#>  [16] "gL(L)"             "IPS(R)"            "OL(R)"            
#>  [19] "SAD"               "SPS(R)"            "NO1(R)"           
#>  [22] "AB(L)"             "a1(R)"             "a'2(R)"           
#>  [25] "ME(R)"             "CRE(L)"            "SPS(L)"           
#>  [28] "gL(R)"             "b1(R)"             "bL(R)"            
#>  [31] "RUB(R)"            "AOTU(R)"           "GA(R)"            
#>  [34] "NO3(R)"            "NO2(R)"            "a3(R)"            
#>  [37] "a'L(L)"            "LAL(-GA)(R)"       "g2(R)"            
#>  [40] "ATL(R)"            "SNP(L)"            "VES(R)"           
#>  [43] "IB"                "FBl3"              "g3(R)"            
#>  [46] "CRE(R)"            "EB"                "CAN(R)"           
#>  [49] "b'1(R)"            "vACA(R)"           "CX"               
#>  [52] "SIP(L)"            "PB(L1)"            "b'L(L)"           
#>  [55] "FBl5"              "PB(R3)"            "AB(R)"            
#>  [58] "aL(L)"             "SLP(R)"            "b'2(R)"           
#>  [61] "PB(L3)"            "lACA(R)"           "CRE(-ROB,-RUB)(R)"
#>  [64] "PB(L9)"            "PB(R2)"            "AOT(R)"           
#>  [67] "CA(L)"             "GF(R)"             "b2(R)"            
#>  [70] "b'L(R)"            "LX(R)"             "VES(L)"           
#>  [73] "AL(R)"             "g4(R)"             "FBl4"             
#>  [76] "LH(R)"             "mALT(L)"           "PRW"              
#>  [79] "g1(R)"             "EBr5"              "ATL(L)"           
#>  [82] "PB(L8)"            "AME(R)"            "NO(L)"            
#>  [85] "PVLP(R)"           "FB"                "EBr3d"            
#>  [88] "LO(R)"             "GC"                "EBr1"             
#>  [91] "VMNP"              "ROB(R)"            "PB(R1)"           
#>  [94] "PB(R6)"            "hemibrain"         "mALT(R)"          
#>  [97] "a'1(R)"            "INP"               "NO2(L)"           
#> [100] "PB(L4)"            "NO1(L)"            "a'L(R)"           
#> [103] "FBl9"              "GOR(R)"            "PENP"             
#> [106] "EBr6"              "EBr2r4"            "PB(R5)"           
#> [109] "NO(R)"             "EBr3pw"            "PB(L7)"           
#> [112] "EPA(L)"            "AMMC"              "BU(R)"            
#> [115] "AL(L)"             "AVLP(R)"           "LOP(R)"           
#> [118] "BU(L)"             "PLP(R)"            "FBl6"             
#> [121] "LAL(R)"            "MB(R)"             "LAL(L)"           
#> [124] "SIP(R)"            "NO"                "LX(L)"            
#> [127] "PB(L6)"            "ICL(R)"            "EBr3am"           
#> [130] "PB(R7)"            "MB(L)"             "PB"               
#> [133] "bL(L)"             "SMP(L)"            "aL(R)"            
#> [136] "PB(R9)"            "dACA(R)"           "FBl2"             
#> [139] "POC"               "RUB(L)"            "GNG"              
#> [142] "a'3(R)"            "GOR(L)"            "EPA(R)"           
#> [145] "FBl8"              "FBl1"              "PB(R4)"           
#> [148] "VLNP(R)"           "a2(R)"             "SCL(R)"           
#> [151] "PB(L2)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method  path                                  
#> [1,] "POST"  "/api/custom/custom"                  
#> [2,] "POST"  "/api/v:ver/npexplorer/autapses"      
#> [3,] "POST"  "/api/npexplorer/distribution"        
#> [4,] "POST"  "/api/raw/keyvalue/key/:instance/:key"
#> [5,] "GET"   "/profile"                            
#> [6,] "PATCH" "/api"                                
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"             
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [5,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [6,] "github.com/labstack/echo.glob..func1"                                                      

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }