Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DA4m(R)"        "g3(R)"             "GA(R)"            
#>   [4] "AL-VA6(L)"         "EPA(L)"            "PED(R)"           
#>   [7] "SPS(L)"            "AME(R)"            "NO3(R)"           
#>  [10] "AL-VC1(R)"         "AL-DL4(R)"         "dACA(R)"          
#>  [13] "FBl9"              "IPS(R)"            "POC"              
#>  [16] "AMMC"              "AL-DL5(L)"         "b1(R)"            
#>  [19] "AL-V(R)"           "VES(L)"            "SAD"              
#>  [22] "ME(R)"             "FBl3"              "AL-DC2(R)"        
#>  [25] "AL-DM5(R)"         "CRE(-RUB)(L)"      "FBl2"             
#>  [28] "a3(R)"             "NO(R)"             "bL(R)"            
#>  [31] "AL-DM5(L)"         "AB(R)"             "LX(L)"            
#>  [34] "AL-DA4l(R)"        "AL-DC1(R)"         "PB(R6)"           
#>  [37] "AL-DC4(L)"         "AL-D(L)"           "AL-DL1(R)"        
#>  [40] "AL-VP1m(R)"        "AL-DM6(L)"         "bL(L)"            
#>  [43] "PB(R4)"            "AL-D(R)"           "CRE(-ROB,-RUB)(R)"
#>  [46] "AL-DM2(L)"         "g5(R)"             "SNP(L)"           
#>  [49] "b'2(R)"            "VMNP"              "gL(L)"            
#>  [52] "SIP(R)"            "LO(R)"             "vACA(R)"          
#>  [55] "AL-DL4(L)"         "a2(R)"             "PRW"              
#>  [58] "WED(R)"            "b'L(R)"            "PB(L1)"           
#>  [61] "a'1(R)"            "IB112"             "AL-VC2(R)"        
#>  [64] "CA(L)"             "mALT(R)"           "EBr2r4"           
#>  [67] "ROB(R)"            "MB(L)"             "PB(L3)"           
#>  [70] "AL-DL2v(R)"        "mALT(L)"           "GOR(L)"           
#>  [73] "PB(L9)"            "AL-VA6(R)"         "AL-VA7l(R)"       
#>  [76] "a'3(R)"            "AL-VP4(R)"         "PLP(R)"           
#>  [79] "gL(R)"             "PB(R2)"            "lACA(R)"          
#>  [82] "b'1(R)"            "SMP(L)"            "aL(L)"            
#>  [85] "AL-DM1(R)"         "AL-VA3(R)"         "PB(R8)"           
#>  [88] "PB(L7)"            "AL-DC4(R)"         "AL-DA3(R)"        
#>  [91] "AL-VM7v(R)"        "FB"                "SAD(-AMMC)"       
#>  [94] "AVLP(R)"           "NO1(L)"            "NO1(R)"           
#>  [97] "AL-VA4(R)"         "AL-VL2a(R)"        "GF(R)"            
#> [100] "PB(R7)"            "AL-DP1l(R)"        "AL-DL3(R)"        
#> [103] "EB"                "NO3(L)"            "AL(R)"            
#> [106] "g4(R)"             "AB(L)"             "MB(+ACA)(R)"      
#> [109] "AL-VA1v(R)"        "AL-VC4(R)"         "AL-VM3(R)"        
#> [112] "RUB(R)"            "SCL(L)"            "CRE(R)"           
#> [115] "AL-VM7v(L)"        "NO2(R)"            "FBl8"             
#> [118] "AL-DL5(R)"         "BU(R)"             "FBl7"             
#> [121] "AL-DA2(R)"         "AL-VL2p(R)"        "AL-VM7d(L)"       
#> [124] "EBr3pw"            "AL-VC3l(R)"        "PB(L5)"           
#> [127] "AL-DC1(L)"         "NO(L)"             "OL(R)"            
#> [130] "b2(R)"             "AL-DM3(L)"         "ICL(L)"           
#> [133] "AL-DP1m(R)"        "AL-VA7m(R)"        "LAL(-GA)(R)"      
#> [136] "EPA(R)"            "CA(R)"             "AL-VA1d(R)"       
#> [139] "AL-VP2(R)"         "AL-VM5d(R)"        "IB"               
#> [142] "EBr6"              "AL-DA3(L)"         "GNG"              
#> [145] "AL-VM5v(R)"        "GC"                "PVLP(R)"          
#> [148] "PB(R3)"            "AL-VM1(R)"         "AL-VP1l(R)"       
#> [151] "BU(L)"             "FLA(R)"            "AL-DM6(R)"        
#> [154] "NO"                "PB(L2)"            "CRE(L)"           
#> [157] "LH(R)"             "SLP(R)"            "AL-VM2(R)"        
#> [160] "AL-VM7d(R)"        "AOTU(R)"           "SCL(R)"           
#> [163] "INP"               "RUB(L)"            "LAL(R)"           
#> [166] "EBr3am"            "aL(R)"             "AL-DM4(L)"        
#> [169] "FBl4"              "LOP(R)"            "LAL(L)"           
#> [172] "AL-DP1m(L)"        "PB(L6)"            "FB-column3"       
#> [175] "b'L(L)"            "AL-VP3(R)"         "AL-VA2(R)"        
#> [178] "AL-VA5(R)"         "NO2(L)"            "AL(L)"            
#> [181] "FBl5"              "a1(R)"             "ICL(R)"           
#> [184] "MB(R)"             "g2(R)"             "AL-DA2(L)"        
#> [187] "PB(R9)"            "SMP(R)"            "hemibrain"        
#> [190] "FBl1"              "ATL(R)"            "a'2(R)"           
#> [193] "GOR(R)"            "CX"                "SNP(R)"           
#> [196] "CAN(R)"            "VLNP(R)"           "PENP"             
#> [199] "AL-DL2d(R)"        "AL-DM1(L)"         "PB(R1)"           
#> [202] "AL-DM3(R)"         "AL-VP1d(R)"        "SIP(L)"           
#> [205] "PB"                "a'L(R)"            "PB(R5)"           
#> [208] "LX(R)"             "PB(L4)"            "AL-DM4(R)"        
#> [211] "SPS(R)"            "AL-DA4m(L)"        "a'L(L)"           
#> [214] "AL-DA1(R)"         "AL-VP5(R)"         "PB(L8)"           
#> [217] "EBr3d"             "VES(R)"            "AL-DM2(R)"        
#> [220] "AL-VC3m(R)"        "AL-VL1(R)"         "g1(R)"            
#> [223] "AL-DC2(L)"         "FBl6"              "EBr5"             
#> [226] "AL-DC3(R)"         "AOT(R)"            "ATL(L)"           
#> [229] "EBr1"              "AL-VC5(R)"         "AL-VM4(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "SPS(L)"            "PB(L6)"            "NO(R)"            
#>   [4] "OL(R)"             "b'1(R)"            "PB(R4)"           
#>   [7] "LOP(R)"            "mALT(L)"           "MB(L)"            
#>  [10] "NO1(R)"            "EB"                "g5(R)"            
#>  [13] "FBl2"              "AMMC"              "b'2(R)"           
#>  [16] "EBr6"              "EBr2r4"            "SLP(R)"           
#>  [19] "AB(L)"             "FLA(R)"            "PB(R2)"           
#>  [22] "AME(R)"            "AL(R)"             "VLNP(R)"          
#>  [25] "a'L(R)"            "AB(R)"             "vACA(R)"          
#>  [28] "b2(R)"             "g2(R)"             "SNP(R)"           
#>  [31] "FBl6"              "PB(R7)"            "LAL(-GA)(R)"      
#>  [34] "a'1(R)"            "ICL(L)"            "INP"              
#>  [37] "SMP(R)"            "CA(L)"             "VES(R)"           
#>  [40] "FBl4"              "FBl9"              "NO2(R)"           
#>  [43] "LX(L)"             "PB(R3)"            "RUB(L)"           
#>  [46] "PB(L3)"            "SIP(R)"            "PB(R1)"           
#>  [49] "gL(R)"             "bL(L)"             "PB(L7)"           
#>  [52] "a1(R)"             "PENP"              "SIP(L)"           
#>  [55] "g1(R)"             "NO"                "CRE(L)"           
#>  [58] "AOTU(R)"           "IPS(R)"            "NO(L)"            
#>  [61] "PB(L4)"            "a'L(L)"            "VES(L)"           
#>  [64] "a2(R)"             "CRE(R)"            "PB(R9)"           
#>  [67] "AVLP(R)"           "ICL(R)"            "FBl1"             
#>  [70] "RUB(R)"            "g3(R)"             "PRW"              
#>  [73] "PB(L9)"            "LH(R)"             "LX(R)"            
#>  [76] "g4(R)"             "PB(L5)"            "SAD"              
#>  [79] "GOR(L)"            "b'L(R)"            "dACA(R)"          
#>  [82] "PB(R8)"            "PB(L1)"            "gL(L)"            
#>  [85] "IB"                "SCL(R)"            "VMNP"             
#>  [88] "EPA(R)"            "ATL(R)"            "PB"               
#>  [91] "GC"                "EBr1"              "MB(R)"            
#>  [94] "BU(L)"             "a'2(R)"            "GNG"              
#>  [97] "LAL(R)"            "FBl8"              "bL(R)"            
#> [100] "AOT(R)"            "hemibrain"         "SPS(R)"           
#> [103] "PED(R)"            "aL(R)"             "SNP(L)"           
#> [106] "mALT(R)"           "PVLP(R)"           "GOR(R)"           
#> [109] "WED(R)"            "b1(R)"             "a'3(R)"           
#> [112] "lACA(R)"           "FBl5"              "EBr3am"           
#> [115] "ATL(L)"            "aL(L)"             "CRE(-ROB,-RUB)(R)"
#> [118] "GF(R)"             "EBr5"              "PB(L8)"           
#> [121] "AL(L)"             "CAN(R)"            "NO3(R)"           
#> [124] "ROB(R)"            "a3(R)"             "PB(R5)"           
#> [127] "BU(R)"             "SAD(-AMMC)"        "MB(+ACA)(R)"      
#> [130] "SCL(L)"            "GA(R)"             "NO1(L)"           
#> [133] "EBr3d"             "FBl3"              "FBl7"             
#> [136] "CA(R)"             "EPA(L)"            "PB(L2)"           
#> [139] "CX"                "LAL(L)"            "FB"               
#> [142] "EBr3pw"            "LO(R)"             "b'L(L)"           
#> [145] "PB(R6)"            "PLP(R)"            "NO3(L)"           
#> [148] "POC"               "SMP(L)"            "ME(R)"            
#> [151] "NO2(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DL5(L)"         "PRW"               "ICL(R)"           
#>   [4] "PB(R6)"            "AL-DL4(L)"         "FBl6"             
#>   [7] "AL-VM1(R)"         "SCL(L)"            "LAL(-GA)(R)"      
#>  [10] "EB"                "aL(L)"             "PB(R7)"           
#>  [13] "AL-DL3(R)"         "PB"                "AL-DC1(L)"        
#>  [16] "PVLP(R)"           "NO2(L)"            "LAL(L)"           
#>  [19] "AL-VC4(R)"         "FBl2"              "WED(R)"           
#>  [22] "SCL(R)"            "CRE(-RUB)(L)"      "EBr3d"            
#>  [25] "a'L(R)"            "AL-VM7v(L)"        "PB(L9)"           
#>  [28] "a3(R)"             "VES(L)"            "AL-DC4(R)"        
#>  [31] "CRE(R)"            "IPS(R)"            "b'L(L)"           
#>  [34] "AL-DC1(R)"         "NO(L)"             "CA(L)"            
#>  [37] "SIP(L)"            "g4(R)"             "b1(R)"            
#>  [40] "vACA(R)"           "CA(R)"             "AL-DA4l(R)"       
#>  [43] "bL(R)"             "SLP(R)"            "AL-DL1(R)"        
#>  [46] "SAD(-AMMC)"        "LAL(R)"            "AL-DP1l(R)"       
#>  [49] "a2(R)"             "a'2(R)"            "AL-V(R)"          
#>  [52] "PLP(R)"            "mALT(L)"           "GF(R)"            
#>  [55] "LH(R)"             "AL-VP5(R)"         "AL-VP4(R)"        
#>  [58] "FBl7"              "AL-VA7l(R)"        "g5(R)"            
#>  [61] "AL-DA3(R)"         "GC"                "FB"               
#>  [64] "AL-DM1(L)"         "CAN(R)"            "mALT(R)"          
#>  [67] "FBl1"              "FBl5"              "PB(L4)"           
#>  [70] "a1(R)"             "PED(R)"            "EPA(R)"           
#>  [73] "GOR(R)"            "SNP(L)"            "PB(R2)"           
#>  [76] "LX(L)"             "NO(R)"             "AL-DM4(L)"        
#>  [79] "FBl3"              "AL-DA2(R)"         "AL-DL4(R)"        
#>  [82] "AL-VA1d(R)"        "AL-VC2(R)"         "AL-VM2(R)"        
#>  [85] "AL-DA2(L)"         "AL-DL5(R)"         "PB(L1)"           
#>  [88] "PB(R8)"            "NO1(L)"            "AL-DC2(R)"        
#>  [91] "POC"               "g1(R)"             "dACA(R)"          
#>  [94] "AL-VP1l(R)"        "AL-DA3(L)"         "AL-VM7d(L)"       
#>  [97] "PB(R3)"            "bL(L)"             "INP"              
#> [100] "AL-VA1v(R)"        "AL-VL1(R)"         "AL-VP1d(R)"       
#> [103] "hemibrain"         "AL-VA3(R)"         "AL-VL2a(R)"       
#> [106] "AL-VM4(R)"         "b'2(R)"            "NO"               
#> [109] "CRE(L)"            "AL-DM1(R)"         "AL-DM3(R)"        
#> [112] "AL-VA6(R)"         "EBr2r4"            "AL-DM2(L)"        
#> [115] "gL(R)"             "AL-VM3(R)"         "AL-VM7d(R)"       
#> [118] "AL-VP1m(R)"        "a'3(R)"            "AL-DC4(L)"        
#> [121] "AL-VM5v(R)"        "AOT(R)"            "AB(L)"            
#> [124] "SNP(R)"            "AL-DC2(L)"         "LO(R)"            
#> [127] "AL-DM4(R)"         "AL-VA5(R)"         "LOP(R)"           
#> [130] "EBr1"              "a'1(R)"            "CRE(-ROB,-RUB)(R)"
#> [133] "PB(L3)"            "b'L(R)"            "FBl8"             
#> [136] "VLNP(R)"           "g3(R)"             "AL-DM6(L)"        
#> [139] "AL-DL2d(R)"        "AL-VM5d(R)"        "AL-VP3(R)"        
#> [142] "CX"                "GOR(L)"            "NO3(R)"           
#> [145] "MB(+ACA)(R)"       "SPS(R)"            "lACA(R)"          
#> [148] "SPS(L)"            "b'1(R)"            "EBr3am"           
#> [151] "a'L(L)"            "AL-VA2(R)"         "gL(L)"            
#> [154] "EBr3pw"            "OL(R)"             "MB(R)"            
#> [157] "AL-DM5(R)"         "AL-VA4(R)"         "AL-VC5(R)"        
#> [160] "BU(R)"             "PB(L6)"            "AL(R)"            
#> [163] "PB(R5)"            "SMP(R)"            "PB(L2)"           
#> [166] "AL-D(L)"           "EPA(L)"            "AL-VC3m(R)"       
#> [169] "AL-VP2(R)"         "BU(L)"             "VMNP"             
#> [172] "AL-DM5(L)"         "SMP(L)"            "EBr6"             
#> [175] "ME(R)"             "PB(R4)"            "PB(L5)"           
#> [178] "GA(R)"             "g2(R)"             "PB(L8)"           
#> [181] "PB(R9)"            "AL-DL2v(R)"        "EBr5"             
#> [184] "VES(R)"            "AB(R)"             "AME(R)"           
#> [187] "AOTU(R)"           "AL-DA4m(L)"        "NO3(L)"           
#> [190] "PB(L7)"            "FBl4"              "AL-VC3l(R)"       
#> [193] "MB(L)"             "LX(R)"             "SIP(R)"           
#> [196] "GNG"               "AL-DC3(R)"         "AL-DA4m(R)"       
#> [199] "ICL(L)"            "RUB(L)"            "AL-DA1(R)"        
#> [202] "NO1(R)"            "AL-VA6(L)"         "AL(L)"            
#> [205] "AVLP(R)"           "NO2(R)"            "AL-DM6(R)"        
#> [208] "AL-VC1(R)"         "FBl9"              "AL-VL2p(R)"       
#> [211] "PENP"              "aL(R)"             "FLA(R)"           
#> [214] "AL-DP1m(L)"        "FB-column3"        "ATL(R)"           
#> [217] "AL-VA7m(R)"        "IB"                "SAD"              
#> [220] "b2(R)"             "AL-DM3(L)"         "AL-DM2(R)"        
#> [223] "AL-DP1m(R)"        "ATL(L)"            "ROB(R)"           
#> [226] "PB(R1)"            "AL-D(R)"           "AL-VM7v(R)"       
#> [229] "RUB(R)"            "AMMC"             
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "MB(R)"             "AL-V(R)"           "PB(R5)"           
#>   [4] "AL-DM6(L)"         "AL-VA5(R)"         "AL-DL4(R)"        
#>   [7] "AL-VC2(R)"         "g2(R)"             "PLP(R)"           
#>  [10] "FBl9"              "AL-VC5(R)"         "a2(R)"            
#>  [13] "LAL(-GA)(R)"       "g5(R)"             "FBl8"             
#>  [16] "EBr3d"             "AL-DA3(R)"         "PB(L9)"           
#>  [19] "RUB(L)"            "b'L(L)"            "a1(R)"            
#>  [22] "AL-DM3(L)"         "NO3(L)"            "AL-DA2(R)"        
#>  [25] "EBr1"              "BU(L)"             "CRE(-RUB)(L)"     
#>  [28] "WED(R)"            "PB(L1)"            "AL-VM4(R)"        
#>  [31] "mALT(L)"           "ATL(L)"            "GF(R)"            
#>  [34] "SAD"               "AL-DM6(R)"         "AMMC"             
#>  [37] "PENP"              "AL-VM7d(L)"        "PB(L7)"           
#>  [40] "CAN(R)"            "AL-DC3(R)"         "AL-DP1l(R)"       
#>  [43] "AB(L)"             "AL-VM7v(L)"        "ROB(R)"           
#>  [46] "AL-DM4(L)"         "PB(R1)"            "b1(R)"            
#>  [49] "AL-DL5(R)"         "AL-VA3(R)"         "LOP(R)"           
#>  [52] "b2(R)"             "gL(L)"             "AL(R)"            
#>  [55] "b'1(R)"            "SCL(L)"            "hemibrain"        
#>  [58] "AL-DC2(L)"         "AL-DC2(R)"         "AL-VA2(R)"        
#>  [61] "AL-VM5v(R)"        "AL-VP1d(R)"        "GOR(L)"           
#>  [64] "PB"                "AL-VL2a(R)"        "SAD(-AMMC)"       
#>  [67] "AL-DA4m(R)"        "CA(R)"             "AL-DC1(R)"        
#>  [70] "AL-DC4(L)"         "FBl1"              "AL-VC4(R)"        
#>  [73] "AL-VM3(R)"         "FBl5"              "PB(R3)"           
#>  [76] "LX(L)"             "AL-VC1(R)"         "GC"               
#>  [79] "a'L(R)"            "AVLP(R)"           "a'3(R)"           
#>  [82] "AL-VA7m(R)"        "AL-DM1(L)"         "AL-DM3(R)"        
#>  [85] "AL-VM1(R)"         "AL-VM2(R)"         "EBr3am"           
#>  [88] "SPS(L)"            "a3(R)"             "a'1(R)"           
#>  [91] "CA(L)"             "PB(R6)"            "g1(R)"            
#>  [94] "AB(R)"             "AL-DL2v(R)"        "SCL(R)"           
#>  [97] "ICL(L)"            "NO2(L)"            "b'L(R)"           
#> [100] "SNP(R)"            "g4(R)"             "AL-VP1m(R)"       
#> [103] "AOTU(R)"           "NO3(R)"            "AL-DL4(L)"        
#> [106] "MB(+ACA)(R)"       "AL-DL3(R)"         "AL-DP1m(R)"       
#> [109] "AL-VM5d(R)"        "AL-VP5(R)"         "vACA(R)"          
#> [112] "PB(L3)"            "SIP(L)"            "AL-DC1(L)"        
#> [115] "CRE(L)"            "AL-VL1(R)"         "AL-VP3(R)"        
#> [118] "PB(R8)"            "AL(L)"             "PVLP(R)"          
#> [121] "AL-VC3l(R)"        "POC"               "CX"               
#> [124] "MB(L)"             "IPS(R)"            "a'2(R)"           
#> [127] "AL-DM2(R)"         "AL-VA1v(R)"        "PB(L5)"           
#> [130] "EBr5"              "LX(R)"             "AL-VA1d(R)"       
#> [133] "AOT(R)"            "PED(R)"            "LO(R)"            
#> [136] "AL-DL5(L)"         "FBl2"              "AL-DA4l(R)"       
#> [139] "SMP(L)"            "VMNP"              "PB(R2)"           
#> [142] "AL-DM1(R)"         "AL-VL2p(R)"        "CRE(-ROB,-RUB)(R)"
#> [145] "EBr2r4"            "BU(R)"             "EPA(L)"           
#> [148] "AL-VP2(R)"         "VES(R)"            "ATL(R)"           
#> [151] "AL-DA1(R)"         "IB"                "EB"               
#> [154] "PB(L4)"            "AL-DC4(R)"         "VLNP(R)"          
#> [157] "PB(R7)"            "FLA(R)"            "AL-DL1(R)"        
#> [160] "AL-VM7v(R)"        "NO1(R)"            "AL-DM5(L)"        
#> [163] "ME(R)"             "AL-VA4(R)"         "AL-VM7d(R)"       
#> [166] "FB"                "dACA(R)"           "NO(L)"            
#> [169] "PB(L8)"            "EPA(R)"            "EBr3pw"           
#> [172] "AL-VC3m(R)"        "b'2(R)"            "lACA(R)"          
#> [175] "AL-VA6(R)"         "AL-VP1l(R)"        "NO2(R)"           
#> [178] "SLP(R)"            "AL-D(R)"           "AL-VP4(R)"        
#> [181] "SIP(R)"            "AL-DM5(R)"         "SMP(R)"           
#> [184] "FBl6"              "AL-VA7l(R)"        "NO(R)"            
#> [187] "FB-column3"        "AL-DL2d(R)"        "RUB(R)"           
#> [190] "AL-DP1m(L)"        "PB(L6)"            "LH(R)"            
#> [193] "INP"               "OL(R)"             "VES(L)"           
#> [196] "AL-DM4(R)"         "FBl4"              "FBl7"             
#> [199] "GOR(R)"            "a'L(L)"            "SPS(R)"           
#> [202] "PB(L2)"            "aL(L)"             "LAL(R)"           
#> [205] "EBr6"              "AL-DA4m(L)"        "AL-DA3(L)"        
#> [208] "AL-D(L)"           "AME(R)"            "LAL(L)"           
#> [211] "PRW"               "AL-VA6(L)"         "AL-DM2(L)"        
#> [214] "GNG"               "mALT(R)"           "CRE(R)"           
#> [217] "PB(R9)"            "g3(R)"             "AL-DA2(L)"        
#> [220] "bL(L)"             "FBl3"              "GA(R)"            
#> [223] "NO1(L)"            "gL(R)"             "bL(R)"            
#> [226] "SNP(L)"            "ICL(R)"            "PB(R4)"           
#> [229] "aL(R)"             "NO"               
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                           
#> [1,] "POST" "/api/npexplorer/roiconnectivity"              
#> [2,] "GET"  "/api/v:ver/npexplorer/celltype/:dataset/:type"
#> [3,] "GET"  "/api/v:ver/dbmeta/database"                   
#> [4,] "POST" "/api/v:ver/npexplorer/autapses"               
#> [5,] "GET"  "/api/dbmeta/instances"                        
#> [6,] "GET"  "/api/npexplorer/neuronmeta"                   
#>      name                                                                                           
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIConnectivity-fm"
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDataInstances-fm"          
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMeta-fm"     

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }