Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0`
#> $`hemibrain:v1.0`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0`$ROIs
#>   [1] "AB(R)"             "LAL(L)"            "LAL(R)"           
#>   [4] "SIP(L)"            "g1(R)"             "ME(R)"            
#>   [7] "ICL(R)"            "NO3(R)"            "NO1(L)"           
#>  [10] "IB"                "VES(R)"            "a2(R)"            
#>  [13] "PRW"               "aL(R)"             "PB(L1)"           
#>  [16] "EBr3d"             "EBr3pw"            "MB(L)"            
#>  [19] "FLA(R)"            "PB(R2)"            "aL(L)"            
#>  [22] "EBr5"              "PB"                "gL(L)"            
#>  [25] "SIP(R)"            "CAN(R)"            "PB(R4)"           
#>  [28] "GOR(L)"            "WED(R)"            "a'1(R)"           
#>  [31] "SNP(L)"            "FBl6"              "PED(R)"           
#>  [34] "AME(R)"            "AMMC"              "BU(L)"            
#>  [37] "a'2(R)"            "FBl1"              "PB(L2)"           
#>  [40] "CRE(-ROB,-RUB)(R)" "SPS(L)"            "a'L(R)"           
#>  [43] "CA(R)"             "CA(L)"             "PB(R7)"           
#>  [46] "SCL(L)"            "a1(R)"             "LOP(R)"           
#>  [49] "SPS(R)"            "lACA(R)"           "LO(R)"            
#>  [52] "GOR(R)"            "VES(L)"            "MB(R)"            
#>  [55] "PB(R6)"            "GC"                "g4(R)"            
#>  [58] "LH(R)"             "b'L(L)"            "PB(L6)"           
#>  [61] "PB(L9)"            "POC"               "MB(+ACA)(R)"      
#>  [64] "FBl9"              "b2(R)"             "NO1(R)"           
#>  [67] "EBr2r4"            "AL(L)"             "PB(R9)"           
#>  [70] "PVLP(R)"           "IPS(R)"            "NO2(R)"           
#>  [73] "b'2(R)"            "AOTU(R)"           "hemibrain"        
#>  [76] "NO"                "FBl2"              "FBl8"             
#>  [79] "mALT(R)"           "PB(L8)"            "BU(R)"            
#>  [82] "SAD(-AMMC)"        "FBl5"              "FB"               
#>  [85] "dACA(R)"           "FBl3"              "NO2(L)"           
#>  [88] "VMNP"              "CRE(L)"            "PB(L4)"           
#>  [91] "OL(R)"             "a'3(R)"            "mALT(L)"          
#>  [94] "GF(R)"             "GA(R)"             "LX(R)"            
#>  [97] "PB(L7)"            "FBl7"              "SAD"              
#> [100] "CRE(R)"            "ICL(L)"            "b1(R)"            
#> [103] "GNG"               "ATL(L)"            "SMP(R)"           
#> [106] "b'1(R)"            "SCL(R)"            "PLP(R)"           
#> [109] "EB"                "RUB(L)"            "NO(L)"            
#> [112] "a'L(L)"            "EBr3am"            "SMP(L)"           
#> [115] "g2(R)"             "EPA(R)"            "NO3(L)"           
#> [118] "LX(L)"             "LAL(-GA)(R)"       "NO(R)"            
#> [121] "PB(L3)"            "EBr1"              "gL(R)"            
#> [124] "AVLP(R)"           "a3(R)"             "bL(R)"            
#> [127] "VLNP(R)"           "AOT(R)"            "vACA(R)"          
#> [130] "SNP(R)"            "PB(R1)"            "PB(R3)"           
#> [133] "CX"                "ROB(R)"            "g5(R)"            
#> [136] "RUB(R)"            "PB(L5)"            "g3(R)"            
#> [139] "EBr6"              "FBl4"              "INP"              
#> [142] "SLP(R)"            "AB(L)"             "AL(R)"            
#> [145] "PB(R5)"            "PENP"              "b'L(R)"           
#> [148] "PB(R8)"            "bL(L)"             "ATL(R)"           
#> [151] "EPA(L)"           
#> 
#> $`hemibrain:v1.0`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method  path                                        
#> [1,] "PATCH" "/api/*"                                    
#> [2,] "PUT"   "/api/*"                                    
#> [3,] "POST"  "/api/v:ver/raw/cypher/cypher"              
#> [4,] "POST"  "/api/v:ver/raw/cypher/transaction/:id/kill"
#> [5,] "GET"   "/api/v:ver/cached/dailytype"               
#> [6,] "GET"   "/api/v:ver/dbmeta/datasets"                
#>      name                                                                
#> [1,] "github.com/labstack/echo.glob..func1"                              
#> [2,] "github.com/labstack/echo.glob..func1"                              
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }