Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0`
#> $`hemibrain:v1.0`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0`$ROIs
#>   [1] "GF(R)"             "SPS(R)"            "GOR(L)"           
#>   [4] "SIP(R)"            "NO"                "FBl7"             
#>   [7] "PB(R2)"            "EBr1"              "BU(L)"            
#>  [10] "BU(R)"             "EBr6"              "PB(L7)"           
#>  [13] "PB(L4)"            "ICL(L)"            "a3(R)"            
#>  [16] "VLNP(R)"           "AMMC"              "PB(L8)"           
#>  [19] "g4(R)"             "CRE(L)"            "gL(R)"            
#>  [22] "b1(R)"             "NO(L)"             "vACA(R)"          
#>  [25] "hemibrain"         "lACA(R)"           "EBr5"             
#>  [28] "GNG"               "PED(R)"            "CAN(R)"           
#>  [31] "PB(R4)"            "b'1(R)"            "EB"               
#>  [34] "a'L(L)"            "AOT(R)"            "LH(R)"            
#>  [37] "AL(R)"             "VMNP"              "LX(R)"            
#>  [40] "gL(L)"             "FBl4"              "FBl9"             
#>  [43] "SMP(R)"            "g3(R)"             "NO1(L)"           
#>  [46] "PB(R9)"            "aL(L)"             "b'L(L)"           
#>  [49] "PB(R5)"            "CRE(R)"            "ATL(R)"           
#>  [52] "LX(L)"             "SLP(R)"            "LOP(R)"           
#>  [55] "aL(R)"             "PB(R1)"            "FBl2"             
#>  [58] "PVLP(R)"           "AVLP(R)"           "NO2(R)"           
#>  [61] "a'3(R)"            "WED(R)"            "NO3(R)"           
#>  [64] "a'L(R)"            "a'1(R)"            "EBr3pw"           
#>  [67] "SNP(L)"            "AB(L)"             "a2(R)"            
#>  [70] "SIP(L)"            "a'2(R)"            "PB(L5)"           
#>  [73] "CRE(-ROB,-RUB)(R)" "AL(L)"             "PB(L9)"           
#>  [76] "g5(R)"             "GOR(R)"            "FBl3"             
#>  [79] "ICL(R)"            "EPA(L)"            "g2(R)"            
#>  [82] "PB(L6)"            "VES(R)"            "FBl1"             
#>  [85] "FBl5"              "CA(L)"             "PENP"             
#>  [88] "FB"                "EBr2r4"            "PRW"              
#>  [91] "NO(R)"             "mALT(R)"           "g1(R)"            
#>  [94] "b2(R)"             "PLP(R)"            "ME(R)"            
#>  [97] "b'2(R)"            "PB(R7)"            "LAL(-GA)(R)"      
#> [100] "AME(R)"            "SCL(L)"            "a1(R)"            
#> [103] "FBl6"              "IB"                "NO2(L)"           
#> [106] "NO1(R)"            "EBr3d"             "bL(L)"            
#> [109] "MB(R)"             "POC"               "LAL(L)"           
#> [112] "GA(R)"             "b'L(R)"            "bL(R)"            
#> [115] "CA(R)"             "RUB(L)"            "RUB(R)"           
#> [118] "GC"                "EPA(R)"            "LO(R)"            
#> [121] "MB(L)"             "MB(+ACA)(R)"       "mALT(L)"          
#> [124] "PB"                "PB(L2)"            "ATL(L)"           
#> [127] "IPS(R)"            "SAD"               "ROB(R)"           
#> [130] "NO3(L)"            "AB(R)"             "INP"              
#> [133] "OL(R)"             "VES(L)"            "CX"               
#> [136] "SCL(R)"            "FLA(R)"            "SNP(R)"           
#> [139] "PB(R3)"            "EBr3am"            "PB(R6)"           
#> [142] "SMP(L)"            "PB(R8)"            "LAL(R)"           
#> [145] "AOTU(R)"           "SPS(L)"            "SAD(-AMMC)"       
#> [148] "FBl8"              "PB(L3)"            "dACA(R)"          
#> [151] "PB(L1)"           
#> 
#> $`hemibrain:v1.0`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method     path                                   
#> [1,] "POST"     "/api/npexplorer/rankedtable"          
#> [2,] "PROPFIND" "/api"                                 
#> [3,] "POST"     "/api/raw/cypher/transaction/:id/kill" 
#> [4,] "POST"     "/api/v:ver/npexplorer/roiconnectivity"
#> [5,] "GET"      "/*"                                   
#> [6,] "POST"     "/api/npexplorer/roiconnectivity"      
#>      name                                                                                           
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getRankedTable-fm"    
#> [2,] "github.com/labstack/echo.glob..func1"                                                         
#> [3,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                       
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [5,] "github.com/labstack/echo.common.static.func1"                                                 
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIConnectivity-fm"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }