Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "GOR(R)"            "VLNP(R)"           "AMMC"             
#>   [4] "PENP"              "g3(R)"             "SPS(L)"           
#>   [7] "PB"                "hemibrain"         "OL(R)"            
#>  [10] "LOP(R)"            "mALT(R)"           "GC"               
#>  [13] "CRE(R)"            "LX(L)"             "a'2(R)"           
#>  [16] "FBl4"              "b'2(R)"            "FLA(R)"           
#>  [19] "CAN(R)"            "PED(R)"            "PB(R3)"           
#>  [22] "AL(R)"             "SMP(R)"            "SAD"              
#>  [25] "PLP(R)"            "gL(L)"             "SCL(L)"           
#>  [28] "ICL(R)"            "GOR(L)"            "aL(R)"            
#>  [31] "EBr5"              "PB(R1)"            "GNG"              
#>  [34] "CX"                "FBl6"              "PB(R6)"           
#>  [37] "NO3(R)"            "SCL(R)"            "EPA(R)"           
#>  [40] "b2(R)"             "NO(L)"             "CRE(-ROB,-RUB)(R)"
#>  [43] "SPS(R)"            "g2(R)"             "NO1(R)"           
#>  [46] "NO3(L)"            "MB(L)"             "g5(R)"            
#>  [49] "FBl1"              "a'L(L)"            "PB(L1)"           
#>  [52] "SAD(-AMMC)"        "a2(R)"             "BU(L)"            
#>  [55] "EBr2r4"            "PB(R9)"            "SIP(R)"           
#>  [58] "a'1(R)"            "PB(L7)"            "EBr3am"           
#>  [61] "ATL(R)"            "a3(R)"             "NO1(L)"           
#>  [64] "EB"                "SNP(L)"            "MB(+ACA)(R)"      
#>  [67] "PB(R5)"            "RUB(R)"            "PB(L8)"           
#>  [70] "EBr3d"             "BU(R)"             "PB(L6)"           
#>  [73] "AME(R)"            "MB(R)"             "ATL(L)"           
#>  [76] "PB(R8)"            "LX(R)"             "ICL(L)"           
#>  [79] "AB(R)"             "PB(L5)"            "WED(R)"           
#>  [82] "VMNP"              "ME(R)"             "PB(L2)"           
#>  [85] "PVLP(R)"           "FBl9"              "EPA(L)"           
#>  [88] "INP"               "GF(R)"             "g1(R)"            
#>  [91] "b'L(R)"            "EBr1"              "IB"               
#>  [94] "VES(R)"            "CA(L)"             "PB(L3)"           
#>  [97] "GA(R)"             "NO"                "LO(R)"            
#> [100] "AOTU(R)"           "dACA(R)"           "RUB(L)"           
#> [103] "bL(R)"             "b'L(L)"            "SLP(R)"           
#> [106] "SIP(L)"            "LAL(-GA)(R)"       "EBr3pw"           
#> [109] "FBl5"              "mALT(L)"           "PRW"              
#> [112] "PB(L9)"            "bL(L)"             "a'3(R)"           
#> [115] "PB(R2)"            "VES(L)"            "gL(R)"            
#> [118] "CRE(L)"            "FBl7"              "LH(R)"            
#> [121] "g4(R)"             "NO(R)"             "AB(L)"            
#> [124] "IPS(R)"            "NO2(L)"            "EBr6"             
#> [127] "SNP(R)"            "ROB(R)"            "CA(R)"            
#> [130] "NO2(R)"            "PB(R4)"            "vACA(R)"          
#> [133] "a1(R)"             "LAL(L)"            "lACA(R)"          
#> [136] "AL(L)"             "FBl8"              "PB(R7)"           
#> [139] "AVLP(R)"           "b'1(R)"            "LAL(R)"           
#> [142] "AOT(R)"            "FB"                "FBl3"             
#> [145] "b1(R)"             "aL(L)"             "POC"              
#> [148] "SMP(L)"            "a'L(R)"            "FBl2"             
#> [151] "PB(L4)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "PB(R6)"            "EPA(R)"            "AL-VA2(R)"        
#>   [4] "AL-VA4(R)"         "a1(R)"             "VES(R)"           
#>   [7] "AL-VC1(R)"         "AL-VM2(R)"         "AOTU(R)"          
#>  [10] "WED(R)"            "SMP(L)"            "AL-DC4(L)"        
#>  [13] "AL-DP1l(R)"        "AL-VA1v(R)"        "AOT(R)"           
#>  [16] "FB"                "gL(L)"             "LAL(R)"           
#>  [19] "AL-VC4(R)"         "AL-VP4(R)"         "AL-DL1(R)"        
#>  [22] "CA(L)"             "g4(R)"             "b1(R)"            
#>  [25] "EPA(L)"            "VES(L)"            "AL-DA4l(R)"       
#>  [28] "EBr5"              "PB(L1)"            "LH(R)"            
#>  [31] "AL-VL1(R)"         "a3(R)"             "AL(R)"            
#>  [34] "a'3(R)"            "GA(R)"             "AL-DL5(L)"        
#>  [37] "SIP(R)"            "AL-VM1(R)"         "RUB(R)"           
#>  [40] "AL-VA7l(R)"        "FLA(R)"            "AL-DA2(L)"        
#>  [43] "SAD(-AMMC)"        "CA(R)"             "AL-VA7m(R)"       
#>  [46] "AB(R)"             "RUB(L)"            "AL-VC5(R)"        
#>  [49] "GF(R)"             "PRW"               "ROB(R)"           
#>  [52] "ATL(R)"            "ICL(R)"            "AL-VM7v(R)"       
#>  [55] "b'1(R)"            "AMMC"              "NO1(R)"           
#>  [58] "AL-DP1m(L)"        "AME(R)"            "AL-DL4(R)"        
#>  [61] "AL-VP1m(R)"        "VMNP"              "FBl5"             
#>  [64] "AL-DP1m(R)"        "g5(R)"             "PB(L4)"           
#>  [67] "AL-DL2v(R)"        "IB"                "EBr3am"           
#>  [70] "FBl3"              "FBl4"              "FBl6"             
#>  [73] "GC"                "b'L(R)"            "dACA(R)"          
#>  [76] "AL-DM3(L)"         "bL(L)"             "OL(R)"            
#>  [79] "AL-VM5v(R)"        "AL-VM3(R)"         "FBl8"             
#>  [82] "EBr3d"             "PB(R3)"            "AL-DM6(R)"        
#>  [85] "AL-VP1d(R)"        "LOP(R)"            "AL-VC3m(R)"       
#>  [88] "AL-VM4(R)"         "NO3(R)"            "a'L(R)"           
#>  [91] "aL(L)"             "NO(L)"             "AL-VC2(R)"        
#>  [94] "LAL(-GA)(R)"       "SNP(L)"            "b2(R)"            
#>  [97] "AL-DL4(L)"         "PB(L2)"            "LAL(L)"           
#> [100] "AL-DM6(L)"         "NO2(R)"            "AL-VP5(R)"        
#> [103] "PB(L5)"            "MB(R)"             "NO2(L)"           
#> [106] "PED(R)"            "AL-VA6(L)"         "PB(L7)"           
#> [109] "AL-DM1(R)"         "AL-VA3(R)"         "PENP"             
#> [112] "ME(R)"             "PLP(R)"            "AL(L)"            
#> [115] "SMP(R)"            "AL-DA1(R)"         "CRE(-ROB,-RUB)(R)"
#> [118] "mALT(L)"           "PB(R7)"            "EB"               
#> [121] "FB-column3"        "AL-DM4(R)"         "AL-DC2(L)"        
#> [124] "bL(R)"             "AL-DC2(R)"         "AL-DC1(L)"        
#> [127] "ICL(L)"            "AL-DA2(R)"         "AL-DL3(R)"        
#> [130] "AL-DL5(R)"         "g3(R)"             "SAD"              
#> [133] "BU(R)"             "FBl2"              "AL-DC3(R)"        
#> [136] "AL-VL2a(R)"        "AL-DA4m(L)"        "AL-D(L)"          
#> [139] "AL-VM7d(L)"        "CRE(L)"            "AL-VA6(R)"        
#> [142] "mALT(R)"           "LX(R)"             "PB(R2)"           
#> [145] "AL-DM5(R)"         "AL-VA1d(R)"        "PB(L3)"           
#> [148] "lACA(R)"           "PB(R9)"            "GOR(R)"           
#> [151] "AL-DM3(R)"         "SPS(R)"            "a2(R)"            
#> [154] "CRE(R)"            "AL-DM5(L)"         "AL-VM5d(R)"       
#> [157] "ATL(L)"            "NO1(L)"            "NO3(L)"           
#> [160] "AL-DM2(R)"         "SIP(L)"            "AL-VP2(R)"        
#> [163] "AB(L)"             "AL-VM7v(L)"        "AL-VL2p(R)"       
#> [166] "SCL(L)"            "BU(L)"             "PB"               
#> [169] "LX(L)"             "AL-DA3(L)"         "IPS(R)"           
#> [172] "AL-DA4m(R)"        "AL-DC1(R)"         "POC"              
#> [175] "hemibrain"         "EBr1"              "PB(R8)"           
#> [178] "EBr2r4"            "MB(L)"             "PB(L9)"           
#> [181] "LO(R)"             "EBr3pw"            "AL-DL2d(R)"       
#> [184] "AL-VA5(R)"         "a'1(R)"            "FBl7"             
#> [187] "CAN(R)"            "b'2(R)"            "aL(R)"            
#> [190] "SPS(L)"            "AL-DA3(R)"         "vACA(R)"          
#> [193] "SCL(R)"            "AL-D(R)"           "SNP(R)"           
#> [196] "PVLP(R)"           "CX"                "PB(R1)"           
#> [199] "FBl9"              "VLNP(R)"           "gL(R)"            
#> [202] "MB(+ACA)(R)"       "AL-VP3(R)"         "a'L(L)"           
#> [205] "GOR(L)"            "a'2(R)"            "AL-DM4(L)"        
#> [208] "PB(L6)"            "NO(R)"             "INP"              
#> [211] "AL-DC4(R)"         "AL-VC3l(R)"        "g2(R)"            
#> [214] "PB(L8)"            "AL-DM1(L)"         "AL-DM2(L)"        
#> [217] "CRE(-RUB)(L)"      "PB(R5)"            "AVLP(R)"          
#> [220] "PB(R4)"            "SLP(R)"            "NO"               
#> [223] "FBl1"              "GNG"               "b'L(L)"           
#> [226] "g1(R)"             "EBr6"              "AL-VM7d(R)"       
#> [229] "AL-V(R)"           "AL-VP1l(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                           
#> [1,] "POST" "/api/raw/cypher/transaction"                  
#> [2,] "GET"  "/api/dbmeta/version"                          
#> [3,] "GET"  "/api/npexplorer/simpleconnections"            
#> [4,] "POST" "/api/v:ver/npexplorer/celltype/:dataset/:type"
#> [5,] "GET"  "/api/npexplorer/rankedtable"                  
#> [6,] "GET"  "/api/npexplorer/distribution"                 
#>      name                                                                                             
#> [1,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                         
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getVersion-fm"                  
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getRankedTable-fm"      
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"     

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }