Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "LO"  "LOP" "ME" 
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "FBl5"              "AL-DM4(R)"         "AL-VP1l(R)"       
#>   [4] "PB(L5)"            "FBl9"              "AL-DA1(R)"        
#>   [7] "AL-VA6(R)"         "AL-VP1d(R)"        "EBr3am"           
#>  [10] "CRE(R)"            "PB(L8)"            "VMNP"             
#>  [13] "g4(R)"             "AL-VL2p(R)"        "AL-VA7l(R)"       
#>  [16] "PB(L7)"            "AL-DL2d(R)"        "GF(R)"            
#>  [19] "gL(L)"             "b'L(L)"            "NO3(L)"           
#>  [22] "IPS(R)"            "GNG"               "AL-DC4(R)"        
#>  [25] "AL-VA1v(R)"        "b2(R)"             "AL-D(L)"          
#>  [28] "NO(L)"             "AL-VM7d(R)"        "PB(L1)"           
#>  [31] "AL-VL2a(R)"        "BU(L)"             "AL-DA4l(R)"       
#>  [34] "CX"                "AL-VP3(R)"         "RUB(L)"           
#>  [37] "AL-DC3(R)"         "AL-DL4(R)"         "AL-VC3l(R)"       
#>  [40] "AL-VM5d(R)"        "PRW"               "AL-DM3(R)"        
#>  [43] "SPS(R)"            "PB(R8)"            "bL(L)"            
#>  [46] "AL-DP1m(R)"        "PB(L4)"            "AL-DL5(R)"        
#>  [49] "GA(R)"             "ATL(R)"            "AL(R)"            
#>  [52] "AL-DL2v(R)"        "PB(L6)"            "AL-VA1d(R)"       
#>  [55] "AL-VA5(R)"         "FBl8"              "POC"              
#>  [58] "AL-DM2(L)"         "SAD(-AMMC)"        "AL-DP1l(R)"       
#>  [61] "PB(R5)"            "SIP(L)"            "EBr2r4"           
#>  [64] "AL(L)"             "AL-DA4m(R)"        "AL-DC1(R)"        
#>  [67] "g3(R)"             "AL-DM6(L)"         "AL-VM7v(L)"       
#>  [70] "ROB(R)"            "CA(R)"             "a2(R)"            
#>  [73] "NO"                "SAD"               "AL-VM1(R)"        
#>  [76] "AOTU(R)"           "PB"                "PB(L9)"           
#>  [79] "PB(R3)"            "LX(L)"             "AL-DL4(L)"        
#>  [82] "AL-DM1(L)"         "AL-DP1m(L)"        "OL(R)"            
#>  [85] "CRE(-ROB,-RUB)(R)" "SNP(L)"            "AL-VC3m(R)"       
#>  [88] "SPS(L)"            "FBl4"              "a'L(L)"           
#>  [91] "LX(R)"             "FB-column3"        "VES(L)"           
#>  [94] "RUB(R)"            "CRE(-RUB)(L)"      "b1(R)"            
#>  [97] "FBl6"              "AME(R)"            "AMMC"             
#> [100] "PED(R)"            "LAL(-GA)(R)"       "NO1(R)"           
#> [103] "PLP(R)"            "EBr3d"             "PB(R6)"           
#> [106] "PB(R1)"            "EPA(L)"            "AL-DM6(R)"        
#> [109] "AL-VM7v(R)"        "SCL(R)"            "mALT(L)"          
#> [112] "lACA(R)"           "aL(R)"             "EBr6"             
#> [115] "a'L(R)"            "AL-DM4(L)"         "SNP(R)"           
#> [118] "PB(R9)"            "EBr3pw"            "AL-DC2(R)"        
#> [121] "NO2(L)"            "MB(L)"             "FB"               
#> [124] "AL-DA4m(L)"        "a'1(R)"            "SLP(R)"           
#> [127] "ME(R)"             "EB"                "BU(R)"            
#> [130] "g5(R)"             "SIP(R)"            "AL-VP5(R)"        
#> [133] "IB112"             "FBl7"              "GOR(R)"           
#> [136] "AL-VM4(R)"         "AL-V(R)"           "dACA(R)"          
#> [139] "AL-VA6(L)"         "FBl2"              "AB(R)"            
#> [142] "AB(L)"             "NO1(L)"            "g2(R)"            
#> [145] "FBl1"              "AL-D(R)"           "a'3(R)"           
#> [148] "LAL(L)"            "AL-VM3(R)"         "ICL(R)"           
#> [151] "CAN(R)"            "AL-VC5(R)"         "PB(L2)"           
#> [154] "LO(R)"             "LH(R)"             "AL-DM2(R)"        
#> [157] "a3(R)"             "AL-VM5v(R)"        "AL-VL1(R)"        
#> [160] "AL-VP2(R)"         "AL-VP1m(R)"        "SMP(L)"           
#> [163] "EBr5"              "FBl3"              "INP"              
#> [166] "AL-DL1(R)"         "IB"                "a1(R)"            
#> [169] "EPA(R)"            "AL-DL5(L)"         "AL-DM1(R)"        
#> [172] "VLNP(R)"           "SMP(R)"            "GC"               
#> [175] "LAL(R)"            "AL-VC1(R)"         "AL-VC2(R)"        
#> [178] "PB(L3)"            "g1(R)"             "bL(R)"            
#> [181] "AL-VM7d(L)"        "gL(R)"             "ICL(L)"           
#> [184] "AL-DL3(R)"         "AOT(R)"            "b'1(R)"           
#> [187] "PENP"              "mALT(R)"           "b'L(R)"           
#> [190] "aL(L)"             "CA(L)"             "PB(R7)"           
#> [193] "AL-DC4(L)"         "MB(+ACA)(R)"       "PB(R2)"           
#> [196] "NO3(R)"            "AL-VA3(R)"         "hemibrain"        
#> [199] "EBr1"              "AL-DA2(L)"         "AL-DC1(L)"        
#> [202] "CRE(L)"            "AVLP(R)"           "AL-VA2(R)"        
#> [205] "a'2(R)"            "AL-DA3(L)"         "AL-VP4(R)"        
#> [208] "SCL(L)"            "NO(R)"             "AL-DA2(R)"        
#> [211] "ATL(L)"            "GOR(L)"            "PB(R4)"           
#> [214] "MB(R)"             "AL-VA7m(R)"        "vACA(R)"          
#> [217] "NO2(R)"            "AL-DA3(R)"         "AL-VM2(R)"        
#> [220] "AL-DC2(L)"         "AL-DM5(L)"         "VES(R)"           
#> [223] "AL-DM5(R)"         "AL-VA4(R)"         "AL-VC4(R)"        
#> [226] "LOP(R)"            "b'2(R)"            "WED(R)"           
#> [229] "FLA(R)"            "AL-DM3(L)"         "PVLP(R)"          
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "SAD"               "WED(R)"            "b2(R)"            
#>   [4] "dACA(R)"           "EBr3am"            "bL(R)"            
#>   [7] "SLP(R)"            "FBl6"              "a'3(R)"           
#>  [10] "PB(R1)"            "PB(L9)"            "VES(R)"           
#>  [13] "a3(R)"             "b'L(L)"            "GOR(L)"           
#>  [16] "g1(R)"             "g2(R)"             "NO3(L)"           
#>  [19] "NO2(R)"            "PB(R6)"            "b'2(R)"           
#>  [22] "PRW"               "FLA(R)"            "SPS(L)"           
#>  [25] "SNP(L)"            "LX(R)"             "SNP(R)"           
#>  [28] "FBl2"              "CX"                "GF(R)"            
#>  [31] "EBr3d"             "ICL(R)"            "PB(L2)"           
#>  [34] "CA(L)"             "SMP(L)"            "SCL(R)"           
#>  [37] "PB(L1)"            "EPA(R)"            "LOP(R)"           
#>  [40] "NO3(R)"            "FB"                "a'2(R)"           
#>  [43] "GNG"               "RUB(R)"            "PLP(R)"           
#>  [46] "g5(R)"             "CAN(R)"            "mALT(L)"          
#>  [49] "PB(R5)"            "LAL(L)"            "PB(L8)"           
#>  [52] "ROB(R)"            "PB(L6)"            "EBr3pw"           
#>  [55] "b1(R)"             "a2(R)"             "EBr6"             
#>  [58] "FBl7"              "FBl8"              "PB(R4)"           
#>  [61] "LAL(R)"            "GA(R)"             "SAD(-AMMC)"       
#>  [64] "LO(R)"             "RUB(L)"            "PB(L5)"           
#>  [67] "a1(R)"             "g3(R)"             "CRE(R)"           
#>  [70] "a'L(R)"            "a'1(R)"            "FBl4"             
#>  [73] "SCL(L)"            "gL(R)"             "vACA(R)"          
#>  [76] "PB(L3)"            "SIP(L)"            "aL(R)"            
#>  [79] "EBr1"              "EBr5"              "VMNP"             
#>  [82] "INP"               "PB(R7)"            "GC"               
#>  [85] "b'L(R)"            "BU(L)"             "PB(R9)"           
#>  [88] "IPS(R)"            "NO(R)"             "FBl9"             
#>  [91] "a'L(L)"            "b'1(R)"            "AMMC"             
#>  [94] "NO2(L)"            "lACA(R)"           "mALT(R)"          
#>  [97] "ME(R)"             "NO"                "ATL(R)"           
#> [100] "OL(R)"             "AL(R)"             "NO1(L)"           
#> [103] "MB(+ACA)(R)"       "FBl3"              "LX(L)"            
#> [106] "LH(R)"             "PB(R2)"            "MB(R)"            
#> [109] "PB(R8)"            "EB"                "gL(L)"            
#> [112] "PVLP(R)"           "EPA(L)"            "LAL(-GA)(R)"      
#> [115] "g4(R)"             "CRE(L)"            "ICL(L)"           
#> [118] "FBl5"              "GOR(R)"            "SPS(R)"           
#> [121] "AL(L)"             "PB(L4)"            "VLNP(R)"          
#> [124] "POC"               "AOT(R)"            "NO1(R)"           
#> [127] "VES(L)"            "IB"                "PENP"             
#> [130] "AOTU(R)"           "hemibrain"         "EBr2r4"           
#> [133] "AVLP(R)"           "PB"                "CA(R)"            
#> [136] "ATL(L)"            "BU(R)"             "FBl1"             
#> [139] "AME(R)"            "NO(L)"             "SMP(R)"           
#> [142] "CRE(-ROB,-RUB)(R)" "SIP(R)"            "bL(L)"            
#> [145] "PB(R3)"            "AB(R)"             "aL(L)"            
#> [148] "PED(R)"            "MB(L)"             "PB(L7)"           
#> [151] "AB(L)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "g2(R)"             "AL-DM5(L)"         "ICL(L)"           
#>   [4] "AL-DA4m(L)"        "FBl5"              "GOR(L)"           
#>   [7] "SAD(-AMMC)"        "VES(R)"            "NO2(R)"           
#>  [10] "AL-DM2(R)"         "AL-VL2a(R)"        "AL-DA3(L)"        
#>  [13] "CRE(L)"            "AVLP(R)"           "LAL(R)"           
#>  [16] "AL-VA1v(R)"        "AL-VP5(R)"         "AL-VP2(R)"        
#>  [19] "ME(R)"             "GF(R)"             "dACA(R)"          
#>  [22] "MB(+ACA)(R)"       "FBl2"              "AL-DM4(R)"        
#>  [25] "AL-DP1l(R)"        "AL-VM5v(R)"        "PRW"              
#>  [28] "NO"                "AL-DM5(R)"         "LX(R)"            
#>  [31] "PB(L9)"            "FBl3"              "AL-VA7l(R)"       
#>  [34] "PB(R6)"            "AL-VM7d(L)"        "AL-VA3(R)"        
#>  [37] "CA(L)"             "aL(R)"             "bL(R)"            
#>  [40] "PVLP(R)"           "AL-VC4(R)"         "EBr5"             
#>  [43] "b1(R)"             "EPA(L)"            "AL-DC2(L)"        
#>  [46] "FBl7"              "GNG"               "GOR(R)"           
#>  [49] "AL-DP1m(L)"        "AL-VC3l(R)"        "a'3(R)"           
#>  [52] "PB(L3)"            "g1(R)"             "NO(R)"            
#>  [55] "aL(L)"             "PB(L4)"            "VES(L)"           
#>  [58] "PED(R)"            "NO3(R)"            "AB(L)"            
#>  [61] "a'L(R)"            "LX(L)"             "SMP(L)"           
#>  [64] "LAL(-GA)(R)"       "AL-DM1(L)"         "AL-DM2(L)"        
#>  [67] "AMMC"              "a1(R)"             "EBr6"             
#>  [70] "AL(L)"             "AL-VP1d(R)"        "PB(L8)"           
#>  [73] "AL-VC1(R)"         "VLNP(R)"           "RUB(R)"           
#>  [76] "EB"                "AL-DC4(L)"         "AL-DA2(R)"        
#>  [79] "mALT(R)"           "SCL(R)"            "AL-VP1l(R)"       
#>  [82] "PB(L5)"            "PB(R9)"            "gL(R)"            
#>  [85] "AL-DA4m(R)"        "AL-D(R)"           "IB"               
#>  [88] "LO(R)"             "PB"                "CRE(-RUB)(L)"     
#>  [91] "CA(R)"             "MB(R)"             "AL-DM3(R)"        
#>  [94] "PLP(R)"            "AL-DM4(L)"         "AL-DL3(R)"        
#>  [97] "NO3(L)"            "b'L(L)"            "AL-DA4l(R)"       
#> [100] "AL-V(R)"           "GC"                "a'2(R)"           
#> [103] "NO(L)"             "AL-DA1(R)"         "AL-DC3(R)"        
#> [106] "b'1(R)"            "NO2(L)"            "NO1(R)"           
#> [109] "AL-DM6(L)"         "FB-column3"        "AL-VP3(R)"        
#> [112] "ATL(L)"            "SPS(R)"            "b'L(R)"           
#> [115] "IPS(R)"            "AL-VM5d(R)"        "AL-VP1m(R)"       
#> [118] "g3(R)"             "PB(L1)"            "EBr2r4"           
#> [121] "BU(R)"             "AL-D(L)"           "PB(R4)"           
#> [124] "SCL(L)"            "a2(R)"             "PB(L6)"           
#> [127] "AB(R)"             "INP"               "AL-DM1(R)"        
#> [130] "PB(R5)"            "SIP(L)"            "BU(L)"            
#> [133] "gL(L)"             "bL(L)"             "AL-DL2d(R)"       
#> [136] "AOT(R)"            "PB(L2)"            "EBr1"             
#> [139] "AL-DC1(L)"         "SLP(R)"            "VMNP"             
#> [142] "SMP(R)"            "PB(R2)"            "AL-VA7m(R)"       
#> [145] "AL-DA2(L)"         "CX"                "CRE(R)"           
#> [148] "SNP(R)"            "AL-DC2(R)"         "AL-DP1m(R)"       
#> [151] "AL-VA1d(R)"        "AOTU(R)"           "AL-DL4(L)"        
#> [154] "ROB(R)"            "AL-VP4(R)"         "POC"              
#> [157] "GA(R)"             "b2(R)"             "ICL(R)"           
#> [160] "AL-VA4(R)"         "FBl6"              "ATL(R)"           
#> [163] "AL-VC3m(R)"        "SNP(L)"            "AL-VM1(R)"        
#> [166] "AL-VM7d(R)"        "LAL(L)"            "hemibrain"        
#> [169] "SPS(L)"            "AL-VM2(R)"         "vACA(R)"          
#> [172] "g4(R)"             "AL-VA5(R)"         "PB(R1)"           
#> [175] "EBr3pw"            "AL-VM3(R)"         "AL-VM4(R)"        
#> [178] "b'2(R)"            "EPA(R)"            "FBl1"             
#> [181] "a3(R)"             "AME(R)"            "AL-DM6(R)"        
#> [184] "FBl9"              "OL(R)"             "AL-VA2(R)"        
#> [187] "mALT(L)"           "NO1(L)"            "AL-DC4(R)"        
#> [190] "AL-VA6(R)"         "AL-VC5(R)"         "AL(R)"            
#> [193] "AL-DL2v(R)"        "PB(R7)"            "AL-DL5(L)"        
#> [196] "a'1(R)"            "a'L(L)"            "WED(R)"           
#> [199] "MB(L)"             "FLA(R)"            "g5(R)"            
#> [202] "AL-VC2(R)"         "PB(R8)"            "EBr3d"            
#> [205] "CRE(-ROB,-RUB)(R)" "PENP"              "FB"               
#> [208] "SAD"               "AL-DM3(L)"         "AL-VA6(L)"        
#> [211] "SIP(R)"            "FBl8"              "LOP(R)"           
#> [214] "LH(R)"             "AL-DL1(R)"         "AL-VL1(R)"        
#> [217] "PB(L7)"            "FBl4"              "CAN(R)"           
#> [220] "PB(R3)"            "RUB(L)"            "AL-DA3(R)"        
#> [223] "AL-DL5(R)"         "AL-VL2p(R)"        "EBr3am"           
#> [226] "lACA(R)"           "AL-VM7v(L)"        "AL-DC1(R)"        
#> [229] "AL-DL4(R)"         "AL-VM7v(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "CA(R)"             "GNG"               "AL-DC2(R)"        
#>   [4] "BU(R)"             "PB(R9)"            "PB(L6)"           
#>   [7] "FBl6"              "b'1(R)"            "a1(R)"            
#>  [10] "a2(R)"             "GC"                "PB(L9)"           
#>  [13] "CAN(R)"            "bL(R)"             "lACA(R)"          
#>  [16] "NO1(R)"            "CRE(L)"            "AL-DA4m(R)"       
#>  [19] "AL-DM2(R)"         "AL-DP1m(R)"        "POC"              
#>  [22] "FBl8"              "PB(L4)"            "IB"               
#>  [25] "b'2(R)"            "AL-DA3(L)"         "AL-DM4(R)"        
#>  [28] "AL-VM5v(R)"        "AOT(R)"            "FBl2"             
#>  [31] "OL(R)"             "AL-VA5(R)"         "AL-DA2(L)"        
#>  [34] "SPS(R)"            "LX(L)"             "AL-VC5(R)"        
#>  [37] "AL-VP3(R)"         "PB(L3)"            "a'2(R)"           
#>  [40] "AL-VA2(R)"         "AL-VA6(R)"         "AL-VP5(R)"        
#>  [43] "AL-VP1d(R)"        "SMP(L)"            "EBr1"             
#>  [46] "PB(L5)"            "AL(L)"             "aL(L)"            
#>  [49] "AL-VA1d(R)"        "PB(R5)"            "CRE(-ROB,-RUB)(R)"
#>  [52] "WED(R)"            "g1(R)"             "ICL(R)"           
#>  [55] "CA(L)"             "GF(R)"             "dACA(R)"          
#>  [58] "g5(R)"             "b1(R)"             "AL-V(R)"          
#>  [61] "SMP(R)"            "gL(L)"             "VES(L)"           
#>  [64] "mALT(L)"           "AL-VC4(R)"         "AL-VM7d(R)"       
#>  [67] "SIP(L)"            "PB(R7)"            "MB(R)"            
#>  [70] "IPS(R)"            "MB(L)"             "EB"               
#>  [73] "FBl7"              "SLP(R)"            "aL(R)"            
#>  [76] "EBr3d"             "AL-DL5(L)"         "AL-DC4(R)"        
#>  [79] "AL-DL5(R)"         "PB(R6)"            "AL-DC1(L)"        
#>  [82] "bL(L)"             "SCL(R)"            "PVLP(R)"          
#>  [85] "AL-DM1(R)"         "NO1(L)"            "EPA(L)"           
#>  [88] "GOR(L)"            "VMNP"              "AL-DA3(R)"        
#>  [91] "a'L(L)"            "AL-DP1l(R)"        "AL-VL2p(R)"       
#>  [94] "g3(R)"             "g4(R)"             "CRE(-RUB)(L)"     
#>  [97] "FBl3"              "AL-VC3m(R)"        "AL-DM6(L)"        
#> [100] "INP"               "SPS(L)"            "AL-DA4m(L)"       
#> [103] "AL-DM4(L)"         "PB(R1)"            "MB(+ACA)(R)"      
#> [106] "LAL(R)"            "AL-DL1(R)"         "RUB(L)"           
#> [109] "FB"                "PB(L8)"            "AL-DM6(R)"        
#> [112] "SCL(L)"            "AL-DC4(L)"         "NO2(R)"           
#> [115] "FBl9"              "b'L(L)"            "CX"               
#> [118] "PB(R8)"            "ROB(R)"            "gL(R)"            
#> [121] "AB(R)"             "EBr5"              "AL-DL2d(R)"       
#> [124] "b'L(R)"            "PB(R4)"            "AL-VL1(R)"        
#> [127] "PB(L1)"            "EPA(R)"            "FB-column3"       
#> [130] "PB(R2)"            "AME(R)"            "AL-VL2a(R)"       
#> [133] "EBr3am"            "LAL(L)"            "LAL(-GA)(R)"      
#> [136] "NO2(L)"            "FLA(R)"            "AVLP(R)"          
#> [139] "AL-VM1(R)"         "RUB(R)"            "NO3(R)"           
#> [142] "BU(L)"             "AL-DC1(R)"         "AL-D(R)"          
#> [145] "AL-VA1v(R)"        "AL-VP4(R)"         "AL-VM7d(L)"       
#> [148] "AL-DL2v(R)"        "hemibrain"         "AL-DC2(L)"        
#> [151] "NO(R)"             "FBl1"              "FBl4"             
#> [154] "PB(R3)"            "a3(R)"             "AL-VC2(R)"        
#> [157] "AL-DL4(R)"         "PLP(R)"            "LX(R)"            
#> [160] "SNP(L)"            "AL-DA1(R)"         "PENP"             
#> [163] "ME(R)"             "AL-VA6(L)"         "ICL(L)"           
#> [166] "VLNP(R)"           "a'3(R)"            "PED(R)"           
#> [169] "AL-DL4(L)"         "a'1(R)"            "NO(L)"            
#> [172] "AL-DA4l(R)"        "AL-DL3(R)"         "a'L(R)"           
#> [175] "ATL(R)"            "AL-VA3(R)"         "EBr2r4"           
#> [178] "AL(R)"             "AL-VM5d(R)"        "AOTU(R)"          
#> [181] "GA(R)"             "AL-D(L)"           "AL-DM5(L)"        
#> [184] "vACA(R)"           "PB(L7)"            "AL-VA7l(R)"       
#> [187] "AL-VM7v(L)"        "AL-VM2(R)"         "LOP(R)"           
#> [190] "SAD"               "PB"                "AL-DP1m(L)"       
#> [193] "SAD(-AMMC)"        "LH(R)"             "AL-VA4(R)"        
#> [196] "AL-VC1(R)"         "AL-VP2(R)"         "AL-VP1l(R)"       
#> [199] "PRW"               "CRE(R)"            "AL-DM3(L)"        
#> [202] "NO3(L)"            "EBr3pw"            "NO"               
#> [205] "GOR(R)"            "AL-VM3(R)"         "EBr6"             
#> [208] "SNP(R)"            "AL-DM2(L)"         "FBl5"             
#> [211] "b2(R)"             "AL-DM1(L)"         "VES(R)"           
#> [214] "AL-DA2(R)"         "AL-VP1m(R)"        "ATL(L)"           
#> [217] "AMMC"              "LO(R)"             "AL-DM5(R)"        
#> [220] "AL-VC3l(R)"        "AL-VM4(R)"         "AL-VM7v(R)"       
#> [223] "mALT(R)"           "g2(R)"             "SIP(R)"           
#> [226] "AL-DC3(R)"         "AL-DM3(R)"         "AL-VA7m(R)"       
#> [229] "AB(L)"             "PB(L2)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                        
#> [1,] "POST" "/api/npexplorer/autapses"                  
#> [2,] "GET"  "/api/skeletons/skeleton/:dataset/:id"      
#> [3,] "HEAD" "/api"                                      
#> [4,] "POST" "/api/v:ver/raw/keyvalue/key/:instance/:key"
#> [5,] "GET"  "/api/v:ver/npexplorer/neuronmeta"          
#> [6,] "POST" "/api/v:ver/npexplorer/neuronmeta"          
#>      name                                                                                    
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/skeletons.masterAPI.getSkeleton-fm"    
#> [3,] "github.com/labstack/echo/v4.glob..func1"                                               
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }