Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "mALT(R)"           "PB(R3)"            "AL-D(R)"          
#>   [4] "AL-DA2(L)"         "AL-DA3(L)"         "ATL(R)"           
#>   [7] "AL-VM7d(R)"        "a'3(R)"            "a1(R)"            
#>  [10] "CA(L)"             "EBr1"              "NO1(R)"           
#>  [13] "PB(L7)"            "AL-VL2p(R)"        "AL-VM2(R)"        
#>  [16] "AL-DA4m(L)"        "FB-column3"        "ICL(L)"           
#>  [19] "LX(L)"             "AL-VP3(R)"         "LAL(-GA)(R)"      
#>  [22] "aL(R)"             "AL-DL5(L)"         "AL-DC2(R)"        
#>  [25] "AL-DL2d(R)"        "PB(L8)"            "FBl7"             
#>  [28] "FBl9"              "AL-VA7l(R)"        "AL-VC2(R)"        
#>  [31] "b'L(R)"            "AL-DM3(L)"         "AL-DM4(R)"        
#>  [34] "AL-VA5(R)"         "AL-VA7m(R)"        "AOTU(R)"          
#>  [37] "AL-VA6(L)"         "RUB(L)"            "AL-DM6(R)"        
#>  [40] "AL-VP5(R)"         "g3(R)"             "AL-DC1(L)"        
#>  [43] "AL-VM7v(L)"        "SLP(R)"            "AL-VA2(R)"        
#>  [46] "SIP(R)"            "GA(R)"             "PB(R7)"           
#>  [49] "AL-DL3(R)"         "AL-VA3(R)"         "AL-DM2(L)"        
#>  [52] "b1(R)"             "PB(R5)"            "AL-DP1m(L)"       
#>  [55] "FBl6"              "EPA(L)"            "bL(R)"            
#>  [58] "AB(L)"             "BU(L)"             "FBl2"             
#>  [61] "PB(R2)"            "AL-DA2(R)"         "AL-VM7v(R)"       
#>  [64] "PB(L3)"            "SPS(L)"            "mALT(L)"          
#>  [67] "EBr5"              "PB(L4)"            "ATL(L)"           
#>  [70] "AL-DM1(L)"         "LO(R)"             "FBl8"             
#>  [73] "ME(R)"             "AL-DM5(R)"         "AL-VP2(R)"        
#>  [76] "NO"                "a2(R)"             "AL-DM5(L)"        
#>  [79] "LH(R)"             "AL-D(L)"           "gL(R)"            
#>  [82] "MB(R)"             "AL-DA3(R)"         "AL-VC1(R)"        
#>  [85] "NO1(L)"            "FLA(R)"            "AVLP(R)"          
#>  [88] "NO(L)"             "PB(L5)"            "PRW"              
#>  [91] "AL-DL5(R)"         "AL(L)"             "CAN(R)"           
#>  [94] "AL-VA6(R)"         "AL-VC3m(R)"        "AMMC"             
#>  [97] "AL-VP4(R)"         "gL(L)"             "EPA(R)"           
#> [100] "PB(L6)"            "AL(R)"             "AL-VL2a(R)"       
#> [103] "PB(R6)"            "PB(L1)"            "PLP(R)"           
#> [106] "AL-DA1(R)"         "CX"                "PB"               
#> [109] "CA(R)"             "ICL(R)"            "AL-VM3(R)"        
#> [112] "FBl1"              "AL-DM3(R)"         "CRE(L)"           
#> [115] "vACA(R)"           "EB"                "a'L(R)"           
#> [118] "POC"               "hemibrain"         "AL-VP1l(R)"       
#> [121] "EBr3d"             "AL-DL4(R)"         "AL-DC2(L)"        
#> [124] "FBl3"              "INP"               "AL-DM1(R)"        
#> [127] "NO3(R)"            "AL-DM4(L)"         "GNG"              
#> [130] "IB"                "GOR(L)"            "b2(R)"            
#> [133] "MB(+ACA)(R)"       "AL-DC4(R)"         "b'1(R)"           
#> [136] "PVLP(R)"           "AL-VM4(R)"         "AL-VM5v(R)"       
#> [139] "a'2(R)"            "AL-DA4l(R)"        "CRE(-ROB,-RUB)(R)"
#> [142] "SMP(L)"            "FBl4"              "SIP(L)"           
#> [145] "IB112"             "b'L(L)"            "AL-DL1(R)"        
#> [148] "AL-VL1(R)"         "PENP"              "PED(R)"           
#> [151] "AL-DM2(R)"         "SCL(R)"            "VLNP(R)"          
#> [154] "b'2(R)"            "SAD"               "AL-DC4(L)"        
#> [157] "AOT(R)"            "lACA(R)"           "EBr2r4"           
#> [160] "SCL(L)"            "AL-DM6(L)"         "AL-VM7d(L)"       
#> [163] "CRE(-RUB)(L)"      "AL-DP1l(R)"        "AL-VP1m(R)"       
#> [166] "RUB(R)"            "SPS(R)"            "WED(R)"           
#> [169] "PB(L9)"            "LAL(R)"            "AL-DL2v(R)"       
#> [172] "NO2(L)"            "AL-VC3l(R)"        "EBr3am"           
#> [175] "LAL(L)"            "SNP(R)"            "FBl5"             
#> [178] "PB(R4)"            "a'L(L)"            "ROB(R)"           
#> [181] "g5(R)"             "AL-VM5d(R)"        "AL-V(R)"          
#> [184] "dACA(R)"           "SNP(L)"            "AL-DC3(R)"        
#> [187] "VES(L)"            "AL-VC5(R)"         "AL-VP1d(R)"       
#> [190] "GF(R)"             "BU(R)"             "PB(R9)"           
#> [193] "EBr3pw"            "OL(R)"             "GC"               
#> [196] "bL(L)"             "SAD(-AMMC)"        "VES(R)"           
#> [199] "LX(R)"             "PB(R1)"            "AL-DC1(R)"        
#> [202] "SMP(R)"            "PB(L2)"            "g2(R)"            
#> [205] "a3(R)"             "AL-DA4m(R)"        "LOP(R)"           
#> [208] "CRE(R)"            "AME(R)"            "AL-DP1m(R)"       
#> [211] "PB(R8)"            "AL-VA1d(R)"        "aL(L)"            
#> [214] "AL-VM1(R)"         "FB"                "g4(R)"            
#> [217] "AL-DL4(L)"         "NO(R)"             "AB(R)"            
#> [220] "g1(R)"             "EBr6"              "NO3(L)"           
#> [223] "AL-VA4(R)"         "MB(L)"             "VMNP"             
#> [226] "IPS(R)"            "GOR(R)"            "AL-VA1v(R)"       
#> [229] "a'1(R)"            "NO2(R)"            "AL-VC4(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "EBr3d"             "MB(+ACA)(R)"       "AL(R)"            
#>   [4] "AOT(R)"            "PB(L8)"            "FBl8"             
#>   [7] "MB(L)"             "LX(R)"             "hemibrain"        
#>  [10] "PRW"               "LAL(-GA)(R)"       "NO(R)"            
#>  [13] "b2(R)"             "SNP(R)"            "a'1(R)"           
#>  [16] "PB(L4)"            "RUB(R)"            "PENP"             
#>  [19] "EBr1"              "NO1(R)"            "ROB(R)"           
#>  [22] "LX(L)"             "GNG"               "bL(R)"            
#>  [25] "ATL(L)"            "SPS(R)"            "EBr5"             
#>  [28] "PB(R8)"            "PB(L9)"            "PVLP(R)"          
#>  [31] "AVLP(R)"           "ICL(R)"            "mALT(R)"          
#>  [34] "CRE(R)"            "PB(L6)"            "MB(R)"            
#>  [37] "AMMC"              "NO2(R)"            "SLP(R)"           
#>  [40] "b'1(R)"            "NO3(R)"            "AL(L)"            
#>  [43] "PB(R3)"            "b'2(R)"            "GOR(L)"           
#>  [46] "AME(R)"            "AB(L)"             "CAN(R)"           
#>  [49] "FB"                "ME(R)"             "NO"               
#>  [52] "LH(R)"             "aL(R)"             "dACA(R)"          
#>  [55] "ICL(L)"            "PB(L7)"            "SNP(L)"           
#>  [58] "EPA(L)"            "CX"                "LAL(L)"           
#>  [61] "AOTU(R)"           "a2(R)"             "PLP(R)"           
#>  [64] "ATL(R)"            "NO(L)"             "CRE(-ROB,-RUB)(R)"
#>  [67] "SMP(L)"            "g3(R)"             "GC"               
#>  [70] "EPA(R)"            "FBl2"              "IPS(R)"           
#>  [73] "GF(R)"             "SCL(R)"            "PB(R1)"           
#>  [76] "OL(R)"             "vACA(R)"           "EBr3am"           
#>  [79] "lACA(R)"           "PB(R7)"            "EB"               
#>  [82] "NO3(L)"            "FBl7"              "IB"               
#>  [85] "PB(L2)"            "WED(R)"            "a'L(R)"           
#>  [88] "FBl5"              "PB(R5)"            "a1(R)"            
#>  [91] "NO2(L)"            "PED(R)"            "g1(R)"            
#>  [94] "PB(R9)"            "GOR(R)"            "a'L(L)"           
#>  [97] "VES(L)"            "mALT(L)"           "BU(R)"            
#> [100] "FBl4"              "LAL(R)"            "POC"              
#> [103] "PB(L5)"            "PB(L3)"            "PB"               
#> [106] "FBl6"              "SMP(R)"            "g4(R)"            
#> [109] "EBr3pw"            "FBl9"              "RUB(L)"           
#> [112] "VES(R)"            "AB(R)"             "b'L(L)"           
#> [115] "SIP(L)"            "g2(R)"             "SIP(R)"           
#> [118] "FBl3"              "CA(R)"             "b'L(R)"           
#> [121] "SPS(L)"            "bL(L)"             "FBl1"             
#> [124] "PB(R2)"            "b1(R)"             "a3(R)"            
#> [127] "aL(L)"             "PB(R6)"            "GA(R)"            
#> [130] "EBr6"              "VMNP"              "a'2(R)"           
#> [133] "gL(R)"             "g5(R)"             "SCL(L)"           
#> [136] "CA(L)"             "PB(L1)"            "FLA(R)"           
#> [139] "gL(L)"             "CRE(L)"            "PB(R4)"           
#> [142] "SAD"               "BU(L)"             "INP"              
#> [145] "LOP(R)"            "NO1(L)"            "EBr2r4"           
#> [148] "SAD(-AMMC)"        "LO(R)"             "VLNP(R)"          
#> [151] "a'3(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DL4(R)"         "vACA(R)"           "a'3(R)"           
#>   [4] "a1(R)"             "PENP"              "SCL(R)"           
#>   [7] "AL-DC2(R)"         "FB-column3"        "AL-VC1(R)"        
#>  [10] "AOT(R)"            "AL-VC4(R)"         "NO(L)"            
#>  [13] "SLP(R)"            "AL-D(R)"           "AL-DC1(R)"        
#>  [16] "mALT(L)"           "IPS(R)"            "PB(L3)"           
#>  [19] "GC"                "BU(R)"             "AL-D(L)"          
#>  [22] "AB(R)"             "GF(R)"             "PB(R8)"           
#>  [25] "PB"                "AL-DM2(R)"         "AL-VM3(R)"        
#>  [28] "LX(L)"             "AL-DL2v(R)"        "AL-VM4(R)"        
#>  [31] "PB(L5)"            "AOTU(R)"           "GOR(L)"           
#>  [34] "bL(L)"             "NO3(R)"            "AL-VC3m(R)"       
#>  [37] "AL-DA2(R)"         "AL-VA4(R)"         "AL-VP2(R)"        
#>  [40] "EBr3am"            "PB(R1)"            "AL-DM6(R)"        
#>  [43] "AL-VM1(R)"         "LOP(R)"            "SAD"              
#>  [46] "NO1(L)"            "AL-DA4m(R)"        "NO2(R)"           
#>  [49] "AL-VL2p(R)"        "AL-DL2d(R)"        "AL-DL3(R)"        
#>  [52] "BU(L)"             "dACA(R)"           "FBl2"             
#>  [55] "MB(+ACA)(R)"       "AL-DC4(R)"         "AL-V(R)"          
#>  [58] "AMMC"              "gL(L)"             "PB(R2)"           
#>  [61] "GNG"               "a2(R)"             "aL(L)"            
#>  [64] "AL-VM5v(R)"        "b'1(R)"            "PB(R9)"           
#>  [67] "ICL(L)"            "AL-DM4(R)"         "mALT(R)"          
#>  [70] "AL-DM1(R)"         "AL-DP1m(R)"        "g4(R)"            
#>  [73] "FBl5"              "FBl7"              "EBr5"             
#>  [76] "SPS(L)"            "ATL(R)"            "AL-VC2(R)"        
#>  [79] "RUB(R)"            "VMNP"              "PB(L7)"           
#>  [82] "AL-DA4l(R)"        "AL-DM4(L)"         "CA(R)"            
#>  [85] "PB(R4)"            "a3(R)"             "CRE(-ROB,-RUB)(R)"
#>  [88] "EBr3d"             "SIP(R)"            "NO"               
#>  [91] "PB(L2)"            "GA(R)"             "lACA(R)"          
#>  [94] "PB(L6)"            "PB(L1)"            "FBl3"             
#>  [97] "FBl6"              "AL-VM7d(R)"        "LX(R)"            
#> [100] "AL-DM5(R)"         "SMP(L)"            "aL(R)"            
#> [103] "g2(R)"             "AL-VM5d(R)"        "AB(L)"            
#> [106] "LAL(L)"            "NO(R)"             "PB(L9)"           
#> [109] "EBr6"              "RUB(L)"            "MB(R)"            
#> [112] "EBr1"              "AL-DL1(R)"         "AL-VP4(R)"        
#> [115] "LAL(-GA)(R)"       "CRE(L)"            "FBl8"             
#> [118] "LO(R)"             "b1(R)"             "AL-VP1d(R)"       
#> [121] "PB(R6)"            "AME(R)"            "NO1(R)"           
#> [124] "SCL(L)"            "AL-DC4(L)"         "CRE(-RUB)(L)"     
#> [127] "SIP(L)"            "AL-VM7d(L)"        "VES(R)"           
#> [130] "g1(R)"             "AL-DA1(R)"         "AL-DA3(R)"        
#> [133] "a'2(R)"            "AL-DC1(L)"         "AL-DM1(L)"        
#> [136] "AVLP(R)"           "AL-VP1m(R)"        "AL(L)"            
#> [139] "ROB(R)"            "AL-DP1l(R)"        "AL-DA4m(L)"       
#> [142] "PB(R7)"            "MB(L)"             "FB"               
#> [145] "AL-VC3l(R)"        "AL-VM2(R)"         "AL-VM7v(R)"       
#> [148] "AL-VP3(R)"         "VLNP(R)"           "SPS(R)"           
#> [151] "a'L(R)"            "PB(L4)"            "AL(R)"            
#> [154] "AL-VA6(R)"         "PB(R5)"            "PED(R)"           
#> [157] "FLA(R)"            "EBr3pw"            "PVLP(R)"          
#> [160] "FBl9"              "POC"               "b'L(R)"           
#> [163] "AL-DM6(L)"         "AL-DP1m(L)"        "SAD(-AMMC)"       
#> [166] "AL-VP5(R)"         "CRE(R)"            "PB(L8)"           
#> [169] "AL-DL5(L)"         "AL-VA1v(R)"        "AL-VA7l(R)"       
#> [172] "NO2(L)"            "b2(R)"             "PLP(R)"           
#> [175] "CAN(R)"            "AL-DM2(L)"         "a'1(R)"           
#> [178] "AL-VA1d(R)"        "AL-VA2(R)"         "AL-VA7m(R)"       
#> [181] "AL-VC5(R)"         "SNP(L)"            "OL(R)"            
#> [184] "AL-DL5(R)"         "FBl1"              "PB(R3)"           
#> [187] "GOR(R)"            "AL-DM3(R)"         "AL-VA3(R)"        
#> [190] "AL-DL4(L)"         "AL-DM3(L)"         "gL(R)"            
#> [193] "LH(R)"             "AL-DC3(R)"         "g3(R)"            
#> [196] "AL-DA3(L)"         "AL-DM5(L)"         "PRW"              
#> [199] "AL-DA2(L)"         "AL-VM7v(L)"        "g5(R)"            
#> [202] "CA(L)"             "SMP(R)"            "NO3(L)"           
#> [205] "ICL(R)"            "b'L(L)"            "CX"               
#> [208] "WED(R)"            "AL-DC2(L)"         "AL-VL1(R)"        
#> [211] "AL-VA5(R)"         "AL-VL2a(R)"        "AL-VP1l(R)"       
#> [214] "ME(R)"             "FBl4"              "EPA(L)"           
#> [217] "INP"               "hemibrain"         "EPA(R)"           
#> [220] "IB"                "EBr2r4"            "bL(R)"            
#> [223] "VES(L)"            "LAL(R)"            "a'L(L)"           
#> [226] "ATL(L)"            "b'2(R)"            "SNP(R)"           
#> [229] "EB"                "AL-VA6(L)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "PB(L5)"            "CA(L)"             "CRE(R)"           
#>   [4] "EBr3d"             "PB(R3)"            "POC"              
#>   [7] "AB(L)"             "FB"                "EBr1"             
#>  [10] "FBl3"              "AL-DC2(R)"         "SPS(L)"           
#>  [13] "b'L(L)"            "AL-DM6(R)"         "AL-VP1m(R)"       
#>  [16] "AL-VA1v(R)"        "SMP(R)"            "EBr3am"           
#>  [19] "LAL(L)"            "VMNP"              "a'2(R)"           
#>  [22] "CRE(L)"            "dACA(R)"           "AL-DM4(R)"        
#>  [25] "PB(R8)"            "PB(L9)"            "AL-DL1(R)"        
#>  [28] "FBl2"              "AL-VP3(R)"         "AL-VA6(L)"        
#>  [31] "FBl4"              "hemibrain"         "AL-VM7d(L)"       
#>  [34] "bL(L)"             "AL-VA7m(R)"        "g2(R)"            
#>  [37] "AL-DM5(R)"         "AL-DP1m(R)"        "AL-DM6(L)"        
#>  [40] "RUB(L)"            "AL-VP4(R)"         "PB(L3)"           
#>  [43] "SLP(R)"            "AL-DM3(R)"         "CRE(-ROB,-RUB)(R)"
#>  [46] "MB(R)"             "AL-VC1(R)"         "AL-VM7v(R)"       
#>  [49] "a'3(R)"            "PB(L4)"            "AL-VM3(R)"        
#>  [52] "AL-VM7d(R)"        "SIP(L)"            "gL(L)"            
#>  [55] "AL-DC1(L)"         "GOR(R)"            "SAD(-AMMC)"       
#>  [58] "PVLP(R)"           "lACA(R)"           "a'L(R)"           
#>  [61] "SIP(R)"            "aL(L)"             "mALT(R)"          
#>  [64] "b2(R)"             "AL-VM7v(L)"        "LO(R)"            
#>  [67] "g4(R)"             "CRE(-RUB)(L)"      "CA(R)"            
#>  [70] "g3(R)"             "NO3(R)"            "EBr3pw"           
#>  [73] "VES(R)"            "ROB(R)"            "ATL(R)"           
#>  [76] "AL-DA4l(R)"        "ATL(L)"            "PED(R)"           
#>  [79] "AL-VC5(R)"         "PB(L2)"            "GF(R)"            
#>  [82] "AL-VP1l(R)"        "SPS(R)"            "PB(L8)"           
#>  [85] "INP"               "AOT(R)"            "mALT(L)"          
#>  [88] "SCL(R)"            "BU(R)"             "LX(L)"            
#>  [91] "AL-VA1d(R)"        "AL-VM5d(R)"        "LOP(R)"           
#>  [94] "LAL(-GA)(R)"       "PLP(R)"            "AL-DM3(L)"        
#>  [97] "PB(R1)"            "AL-DC1(R)"         "AL-DM5(L)"        
#> [100] "AME(R)"            "a3(R)"             "AL-DA1(R)"        
#> [103] "PB(L7)"            "CAN(R)"            "VES(L)"           
#> [106] "AL-VC4(R)"         "AL-VM5v(R)"        "GA(R)"            
#> [109] "LH(R)"             "b1(R)"             "IPS(R)"           
#> [112] "b'L(R)"            "PB(L6)"            "FBl8"             
#> [115] "AL-VC2(R)"         "AL-VL1(R)"         "AL-VP5(R)"        
#> [118] "WED(R)"            "NO1(L)"            "PB"               
#> [121] "AL-DA3(L)"         "AL-DC4(L)"         "AL-D(L)"          
#> [124] "GNG"               "AL-DP1l(R)"        "aL(R)"            
#> [127] "g5(R)"             "PB(R4)"            "AL-DL4(R)"        
#> [130] "FBl6"              "EBr2r4"            "AL-DC2(L)"        
#> [133] "gL(R)"             "NO3(L)"            "SMP(L)"           
#> [136] "LX(R)"             "FB-column3"        "FBl5"             
#> [139] "NO1(R)"            "AL-DL4(L)"         "NO"               
#> [142] "AL-VA6(R)"         "AOTU(R)"           "b'1(R)"           
#> [145] "AL(L)"             "NO(R)"             "AL-VA4(R)"        
#> [148] "vACA(R)"           "SAD"               "EBr6"             
#> [151] "EPA(R)"            "PB(L1)"            "AL-DL2v(R)"       
#> [154] "AL-VM1(R)"         "MB(+ACA)(R)"       "AL-DM2(R)"        
#> [157] "AMMC"              "PB(R6)"            "AL-DA4m(L)"       
#> [160] "AL-VA7l(R)"        "RUB(R)"            "IB"               
#> [163] "AVLP(R)"           "AL-DC3(R)"         "AL-DC4(R)"        
#> [166] "AL-VA5(R)"         "b'2(R)"            "AL-DM4(L)"        
#> [169] "NO(L)"             "PB(R5)"            "a2(R)"            
#> [172] "AL-VL2a(R)"        "AL-VL2p(R)"        "AL-DA2(R)"        
#> [175] "PB(R7)"            "AL-DL5(L)"         "ICL(R)"           
#> [178] "NO2(R)"            "ME(R)"             "AL-DP1m(L)"       
#> [181] "AL-DL5(R)"         "NO2(L)"            "PB(R2)"           
#> [184] "OL(R)"             "AL-DL2d(R)"        "AL-VP2(R)"        
#> [187] "AL-VC3m(R)"        "AL-VM2(R)"         "a1(R)"            
#> [190] "FLA(R)"            "AL-DM2(L)"         "AB(R)"            
#> [193] "AL-VC3l(R)"        "GOR(L)"            "PB(R9)"           
#> [196] "GC"                "SNP(L)"            "AL-VA2(R)"        
#> [199] "a'1(R)"            "FBl9"              "AL-DA4m(R)"       
#> [202] "SCL(L)"            "PENP"              "EB"               
#> [205] "bL(R)"             "AL-DL3(R)"         "g1(R)"            
#> [208] "VLNP(R)"           "AL-V(R)"           "BU(L)"            
#> [211] "ICL(L)"            "FBl7"              "AL-VA3(R)"        
#> [214] "CX"                "PRW"               "AL-DA2(L)"        
#> [217] "AL-DM1(R)"         "AL-VP1d(R)"        "SNP(R)"           
#> [220] "a'L(L)"            "AL-D(R)"           "MB(L)"            
#> [223] "AL-DA3(R)"         "AL-DM1(L)"         "LAL(R)"           
#> [226] "AL(R)"             "EBr5"              "FBl1"             
#> [229] "EPA(L)"            "AL-VM4(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                        
#> [1,] "GET"  "/profile"                                  
#> [2,] "GET"  "/token"                                    
#> [3,] "POST" "/api/v:ver/npexplorer/neuronmeta"          
#> [4,] "POST" "/api/v:ver/raw/cypher/transaction/:id/kill"
#> [5,] "POST" "/api/npexplorer/distribution"              
#> [6,] "GET"  "/api/npexplorer/findneurons"               
#>      name                                                                                        
#> [1,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [2,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getFindNeurons-fm" 

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }