Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "GOR(R)"            "NO3(R)"            "LX(L)"            
#>   [4] "LH(R)"             "SNP(L)"            "FBl1"             
#>   [7] "WED(R)"            "NO"                "PB(R7)"           
#>  [10] "vACA(R)"           "a'L(R)"            "EBr3pw"           
#>  [13] "LO(R)"             "FBl6"              "b'L(R)"           
#>  [16] "SNP(R)"            "FBl4"              "CA(R)"            
#>  [19] "ATL(R)"            "LAL(R)"            "AMMC"             
#>  [22] "dACA(R)"           "a'1(R)"            "ICL(R)"           
#>  [25] "CAN(R)"            "PB(R6)"            "NO3(L)"           
#>  [28] "ICL(L)"            "CX"                "BU(L)"            
#>  [31] "PB(L1)"            "AB(R)"             "AL(R)"            
#>  [34] "PB(L2)"            "SPS(R)"            "VMNP"             
#>  [37] "gL(R)"             "g2(R)"             "a'2(R)"           
#>  [40] "PB(L6)"            "FBl9"              "PB(R2)"           
#>  [43] "aL(L)"             "GOR(L)"            "mALT(R)"          
#>  [46] "AB(L)"             "a1(R)"             "INP"              
#>  [49] "VLNP(R)"           "RUB(R)"            "SCL(L)"           
#>  [52] "SIP(R)"            "PB(R3)"            "FBl5"             
#>  [55] "hemibrain"         "MB(+ACA)(R)"       "FB"               
#>  [58] "EB"                "EBr3am"            "PENP"             
#>  [61] "PED(R)"            "LAL(-GA)(R)"       "EBr5"             
#>  [64] "PB"                "bL(L)"             "FBl3"             
#>  [67] "VES(L)"            "b'2(R)"            "AOTU(R)"          
#>  [70] "AOT(R)"            "LOP(R)"            "PB(R5)"           
#>  [73] "BU(R)"             "RUB(L)"            "CRE(L)"           
#>  [76] "OL(R)"             "LAL(L)"            "NO1(R)"           
#>  [79] "ATL(L)"            "EBr3d"             "b1(R)"            
#>  [82] "AME(R)"            "SMP(R)"            "CA(L)"            
#>  [85] "SCL(R)"            "SPS(L)"            "AL(L)"            
#>  [88] "NO2(R)"            "EPA(L)"            "a'3(R)"           
#>  [91] "NO1(L)"            "PB(L8)"            "g3(R)"            
#>  [94] "SAD"               "SAD(-AMMC)"        "a3(R)"            
#>  [97] "g1(R)"             "aL(R)"             "EPA(R)"           
#> [100] "PB(R9)"            "ROB(R)"            "PB(L9)"           
#> [103] "PB(L7)"            "PB(R4)"            "EBr1"             
#> [106] "gL(L)"             "PB(L5)"            "SLP(R)"           
#> [109] "IB"                "IPS(R)"            "POC"              
#> [112] "b'1(R)"            "NO2(L)"            "EBr6"             
#> [115] "FBl8"              "FBl2"              "FBl7"             
#> [118] "mALT(L)"           "b2(R)"             "LX(R)"            
#> [121] "GNG"               "SIP(L)"            "g5(R)"            
#> [124] "CRE(R)"            "a'L(L)"            "GA(R)"            
#> [127] "PRW"               "lACA(R)"           "MB(L)"            
#> [130] "ME(R)"             "g4(R)"             "GF(R)"            
#> [133] "SMP(L)"            "CRE(-ROB,-RUB)(R)" "AVLP(R)"          
#> [136] "b'L(L)"            "GC"                "FLA(R)"           
#> [139] "PB(R1)"            "NO(R)"             "PB(L3)"           
#> [142] "a2(R)"             "PVLP(R)"           "PLP(R)"           
#> [145] "VES(R)"            "bL(R)"             "NO(L)"            
#> [148] "PB(L4)"            "MB(R)"             "PB(R8)"           
#> [151] "EBr2r4"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AVLP(R)"           "AL-VM5d(R)"        "vACA(R)"          
#>   [4] "g1(R)"             "ICL(L)"            "AL-DA1(R)"        
#>   [7] "AL-VP4(R)"         "mALT(R)"           "FBl3"             
#>  [10] "AL-DM5(R)"         "AL-VC5(R)"         "PRW"              
#>  [13] "EBr3d"             "FBl5"              "AL-VC3m(R)"       
#>  [16] "a'2(R)"            "g5(R)"             "LH(R)"            
#>  [19] "dACA(R)"           "PB"                "FBl1"             
#>  [22] "FBl8"              "GC"                "VMNP"             
#>  [25] "a3(R)"             "AL-DL5(R)"         "EBr3am"           
#>  [28] "AL-DA4m(R)"        "SAD"               "AL-D(L)"          
#>  [31] "AL-VL2p(R)"        "AL-VM7v(R)"        "GOR(L)"           
#>  [34] "b'L(R)"            "PB(L8)"            "GOR(R)"           
#>  [37] "SNP(R)"            "AL(L)"             "AL-DM2(L)"        
#>  [40] "AOT(R)"            "g2(R)"             "PVLP(R)"          
#>  [43] "FB"                "AL-DA2(R)"         "a'3(R)"           
#>  [46] "g3(R)"             "PB(L7)"            "LO(R)"            
#>  [49] "GNG"               "AL-VL2a(R)"        "SMP(L)"           
#>  [52] "AL(R)"             "AL-V(R)"           "LOP(R)"           
#>  [55] "PB(L5)"            "a1(R)"             "AL-DA3(L)"        
#>  [58] "AL-DL1(R)"         "AL-DL4(R)"         "AL-DM4(R)"        
#>  [61] "AL-VL1(R)"         "PB(R7)"            "EBr3pw"           
#>  [64] "PB(R4)"            "FLA(R)"            "a'1(R)"           
#>  [67] "CA(R)"             "OL(R)"             "PB(L3)"           
#>  [70] "BU(L)"             "AL-VM7d(R)"        "b2(R)"            
#>  [73] "MB(L)"             "AL-DM1(L)"         "AL-VA5(R)"        
#>  [76] "AL-VP5(R)"         "AL-VP1l(R)"        "LX(R)"            
#>  [79] "SAD(-AMMC)"        "FBl6"              "FBl9"             
#>  [82] "b'L(L)"            "EBr2r4"            "LX(L)"            
#>  [85] "CX"                "AL-DC2(L)"         "IPS(R)"           
#>  [88] "EPA(L)"            "aL(R)"             "AL-DA3(R)"        
#>  [91] "AL-DA4l(R)"        "AL-DC1(R)"         "mALT(L)"          
#>  [94] "AL-DM5(L)"         "SIP(R)"            "AME(R)"           
#>  [97] "AL-VM4(R)"         "SCL(R)"            "AL-DC3(R)"        
#> [100] "CA(L)"             "LAL(-GA)(R)"       "NO1(L)"           
#> [103] "a'L(R)"            "PB(R3)"            "AL-DL3(R)"        
#> [106] "AL-VM3(R)"         "PB(R5)"            "AMMC"             
#> [109] "bL(R)"             "PENP"              "LAL(L)"           
#> [112] "NO3(L)"            "AL-DC4(R)"         "AL-DM1(R)"        
#> [115] "CRE(-ROB,-RUB)(R)" "SPS(L)"            "AL-DL2v(R)"       
#> [118] "AL-VA3(R)"         "NO3(R)"            "NO"               
#> [121] "AL-VA2(R)"         "AL-VP1d(R)"        "AL-DP1m(L)"       
#> [124] "b1(R)"             "RUB(L)"            "AL-DM2(R)"        
#> [127] "AL-VC2(R)"         "AL-VC3l(R)"        "hemibrain"        
#> [130] "ME(R)"             "AL-DC4(L)"         "AL-DC2(R)"        
#> [133] "AL-DP1l(R)"        "AL-VA7l(R)"        "AL-VM1(R)"        
#> [136] "b'2(R)"            "EBr6"              "PB(R8)"           
#> [139] "PB(R9)"            "AL-VM5v(R)"        "b'1(R)"           
#> [142] "SCL(L)"            "ATL(L)"            "ICL(R)"           
#> [145] "ATL(R)"            "CAN(R)"            "AL-VC4(R)"        
#> [148] "ROB(R)"            "SNP(L)"            "PB(L2)"           
#> [151] "AL-DA4m(L)"        "FBl2"              "AL-D(R)"          
#> [154] "EPA(R)"            "AL-DA2(L)"         "LAL(R)"           
#> [157] "SLP(R)"            "AB(L)"             "GA(R)"            
#> [160] "PED(R)"            "PB(R1)"            "MB(+ACA)(R)"      
#> [163] "AL-VA4(R)"         "PB(L6)"            "aL(L)"            
#> [166] "AL-VP1m(R)"        "EB"                "g4(R)"            
#> [169] "NO(R)"             "PB(L9)"            "CRE(-RUB)(L)"     
#> [172] "PB(R2)"            "AL-VA1v(R)"        "POC"              
#> [175] "IB"                "EBr1"              "EBr5"             
#> [178] "INP"               "PB(L4)"            "gL(L)"            
#> [181] "AL-DM3(L)"         "NO2(L)"            "NO1(R)"           
#> [184] "VES(R)"            "FBl7"              "AL-VP2(R)"        
#> [187] "WED(R)"            "AL-VM7d(L)"        "NO2(R)"           
#> [190] "AL-DP1m(R)"        "AL-VA7m(R)"        "AL-VP3(R)"        
#> [193] "VLNP(R)"           "GF(R)"             "AL-DC1(L)"        
#> [196] "AL-DL5(L)"         "AL-DL2d(R)"        "AL-DM3(R)"        
#> [199] "PB(R6)"            "gL(R)"             "FB-column3"       
#> [202] "VES(L)"            "AL-VA6(R)"         "AL-DM6(R)"        
#> [205] "a2(R)"             "PLP(R)"            "MB(R)"            
#> [208] "AL-VA1d(R)"        "CRE(R)"            "NO(L)"            
#> [211] "SPS(R)"            "AL-DM4(L)"         "bL(L)"            
#> [214] "AL-VM2(R)"         "AL-DM6(L)"         "AL-VA6(L)"        
#> [217] "SMP(R)"            "RUB(R)"            "SIP(L)"           
#> [220] "PB(L1)"            "a'L(L)"            "AL-VC1(R)"        
#> [223] "AL-VM7v(L)"        "AB(R)"             "AOTU(R)"          
#> [226] "CRE(L)"            "lACA(R)"           "BU(R)"            
#> [229] "AL-DL4(L)"         "FBl4"             
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                    
#> [1,] "POST" "/api/npexplorer/roisinneuron"          
#> [2,] "POST" "/api/raw/cypher/transaction/:id/commit"
#> [3,] "GET"  "/token"                                
#> [4,] "GET"  "/api/v:ver/npexplorer/neuronmeta"      
#> [5,] "GET"  "/api/cached/dailytype"                 
#> [6,] "GET"  "/oauth2callback"                       
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"
#> [2,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [3,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getDailyType-fm"          
#> [6,] "github.com/janelia-flyem/echo-secure.oauthCallbackHandler"                                 

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }