Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "FBl2"              "PB(R2)"            "AL-VL1(R)"        
#>   [4] "AL-VM1(R)"         "AL-VP4(R)"         "ATL(R)"           
#>   [7] "RUB(L)"            "SLP(R)"            "AL-DM3(R)"        
#>  [10] "AL-VL2p(R)"        "b'1(R)"            "AL-DA2(L)"        
#>  [13] "g1(R)"             "CRE(R)"            "PLP(R)"           
#>  [16] "AL-VC5(R)"         "AL-VM4(R)"         "SPS(L)"           
#>  [19] "AL-DM4(L)"         "SCL(L)"            "AMMC"             
#>  [22] "g2(R)"             "AL-DA3(L)"         "AL-DC4(R)"        
#>  [25] "RUB(R)"            "NO2(L)"            "PB(R6)"           
#>  [28] "AL-VA6(R)"         "NO(R)"             "NO1(L)"           
#>  [31] "EBr1"              "MB(R)"             "AL-VA7l(R)"       
#>  [34] "AL-VL2a(R)"        "IB"                "PB(L1)"           
#>  [37] "AL-VA6(L)"         "AL-VM5d(R)"        "LH(R)"            
#>  [40] "a3(R)"             "AL-V(R)"           "PVLP(R)"          
#>  [43] "FBl1"              "FBl7"              "mALT(L)"          
#>  [46] "EBr3d"             "gL(L)"             "AL-VM2(R)"        
#>  [49] "AL-VM7v(R)"        "AB(L)"             "AOTU(R)"          
#>  [52] "SNP(L)"            "AL-DL1(R)"         "AL-VM7d(R)"       
#>  [55] "PRW"               "aL(R)"             "AL-DC2(L)"        
#>  [58] "SIP(R)"            "b'L(L)"            "AL-DA2(R)"        
#>  [61] "a'3(R)"            "LX(R)"             "AL-DM5(L)"        
#>  [64] "PB(L6)"            "AL-VC1(R)"         "AL-VP2(R)"        
#>  [67] "AOT(R)"            "CX"                "SCL(R)"           
#>  [70] "PB(L4)"            "AL-DC2(R)"         "AL-VA3(R)"        
#>  [73] "AL-DA4l(R)"        "AL-VC4(R)"         "ATL(L)"           
#>  [76] "g5(R)"             "AL-DC1(R)"         "AL-DP1l(R)"       
#>  [79] "BU(L)"             "EBr2r4"            "a'L(R)"           
#>  [82] "NO(L)"             "AL-VC3l(R)"        "CRE(-ROB,-RUB)(R)"
#>  [85] "GOR(L)"            "EBr5"              "EPA(R)"           
#>  [88] "ICL(L)"            "VES(R)"            "AL-VA2(R)"        
#>  [91] "b2(R)"             "AL-DM6(L)"         "NO3(L)"           
#>  [94] "LAL(R)"            "AL-DM1(L)"         "gL(R)"            
#>  [97] "SPS(R)"            "CRE(-RUB)(L)"      "PB(L7)"           
#> [100] "AL-DM4(R)"         "GNG"               "SMP(L)"           
#> [103] "PB(L8)"            "PB(R8)"            "AL-DC1(L)"        
#> [106] "PB(R9)"            "EBr3pw"            "AME(R)"           
#> [109] "AL-DL2v(R)"        "PB(L3)"            "AL-VM7v(L)"       
#> [112] "VES(L)"            "AL-VA1d(R)"        "CA(L)"            
#> [115] "dACA(R)"           "AL-DM1(R)"         "GA(R)"            
#> [118] "EBr6"              "AL(R)"             "AL-VP1d(R)"       
#> [121] "WED(R)"            "NO3(R)"            "VMNP"             
#> [124] "a1(R)"             "hemibrain"         "SNP(R)"           
#> [127] "AL-DM2(L)"         "FBl6"              "OL(R)"            
#> [130] "SIP(L)"            "AL-DC4(L)"         "AL-DM3(L)"        
#> [133] "FBl8"              "a2(R)"             "PB(L9)"           
#> [136] "PED(R)"            "BU(R)"             "PB(R1)"           
#> [139] "AL-DM6(R)"         "g3(R)"             "PB(R4)"           
#> [142] "GOR(R)"            "AL-DC3(R)"         "MB(L)"            
#> [145] "FB"                "ME(R)"             "AL-DP1m(L)"       
#> [148] "bL(L)"             "aL(L)"             "mALT(R)"          
#> [151] "AL-VM7d(L)"        "CA(R)"             "AB(R)"            
#> [154] "SAD(-AMMC)"        "a'L(L)"            "AL-DA4m(L)"       
#> [157] "FBl3"              "SMP(R)"            "NO"               
#> [160] "AL-DL5(R)"         "AL-DM5(R)"         "SAD"              
#> [163] "PB(R5)"            "FB-column3"        "b1(R)"            
#> [166] "LAL(L)"            "LAL(-GA)(R)"       "CRE(L)"           
#> [169] "ICL(R)"            "AL-D(L)"           "NO2(R)"           
#> [172] "AL-VP1m(R)"        "EBr3am"            "PB(R7)"           
#> [175] "AL-VP3(R)"         "LX(L)"             "AL-DL4(R)"        
#> [178] "AL-VC2(R)"         "AL-VP5(R)"         "AL-VA1v(R)"       
#> [181] "VLNP(R)"           "AL-DL4(L)"         "AL-DP1m(R)"       
#> [184] "AL-VP1l(R)"        "POC"               "PB(L5)"           
#> [187] "a'2(R)"            "IB112"             "PB(L2)"           
#> [190] "vACA(R)"           "GC"                "PB"               
#> [193] "AL-DA1(R)"         "PENP"              "b'L(R)"           
#> [196] "a'1(R)"            "FBl4"              "CAN(R)"           
#> [199] "AL-DL3(R)"         "AVLP(R)"           "EPA(L)"           
#> [202] "INP"               "bL(R)"             "AL-DA4m(R)"       
#> [205] "AL-DA3(R)"         "AL-VM3(R)"         "AL-VM5v(R)"       
#> [208] "AL(L)"             "AL-DL5(L)"         "ROB(R)"           
#> [211] "AL-DM2(R)"         "AL-VA7m(R)"        "AL-VA5(R)"        
#> [214] "FBl5"              "AL-DL2d(R)"        "AL-D(R)"          
#> [217] "LOP(R)"            "b'2(R)"            "EB"               
#> [220] "FBl9"              "GF(R)"             "FLA(R)"           
#> [223] "LO(R)"             "AL-VC3m(R)"        "lACA(R)"          
#> [226] "NO1(R)"            "g4(R)"             "MB(+ACA)(R)"      
#> [229] "PB(R3)"            "IPS(R)"            "AL-VA4(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "lACA(R)"           "SIP(R)"            "g5(R)"            
#>   [4] "LX(L)"             "EBr3d"             "FBl4"             
#>   [7] "AME(R)"            "aL(L)"             "MB(L)"            
#>  [10] "PB(R1)"            "PB(L9)"            "GF(R)"            
#>  [13] "EBr6"              "IPS(R)"            "PB(R4)"           
#>  [16] "PB(R5)"            "bL(L)"             "GNG"              
#>  [19] "SMP(L)"            "GOR(L)"            "NO1(L)"           
#>  [22] "AVLP(R)"           "AL(R)"             "GC"               
#>  [25] "LX(R)"             "a'2(R)"            "EB"               
#>  [28] "g4(R)"             "SNP(L)"            "VES(L)"           
#>  [31] "hemibrain"         "aL(R)"             "PB(R8)"           
#>  [34] "ME(R)"             "RUB(L)"            "b'2(R)"           
#>  [37] "LAL(L)"            "SAD(-AMMC)"        "MB(+ACA)(R)"      
#>  [40] "PB(L4)"            "g2(R)"             "NO"               
#>  [43] "VES(R)"            "a'L(L)"            "CA(L)"            
#>  [46] "PVLP(R)"           "FBl9"              "AB(R)"            
#>  [49] "AOT(R)"            "PB(L3)"            "GA(R)"            
#>  [52] "mALT(R)"           "dACA(R)"           "AL(L)"            
#>  [55] "NO3(L)"            "FBl3"              "b'1(R)"           
#>  [58] "EBr3pw"            "FBl2"              "NO2(R)"           
#>  [61] "bL(R)"             "RUB(R)"            "PB(L2)"           
#>  [64] "CRE(-ROB,-RUB)(R)" "a1(R)"             "NO2(L)"           
#>  [67] "gL(L)"             "ICL(R)"            "b'L(L)"           
#>  [70] "IB"                "PB(R7)"            "LO(R)"            
#>  [73] "a3(R)"             "LAL(R)"            "MB(R)"            
#>  [76] "POC"               "ATL(L)"            "CX"               
#>  [79] "PB(R6)"            "CRE(R)"            "PB(R9)"           
#>  [82] "ATL(R)"            "b1(R)"             "EPA(L)"           
#>  [85] "AB(L)"             "NO3(R)"            "EBr5"             
#>  [88] "SPS(L)"            "EBr2r4"            "ROB(R)"           
#>  [91] "OL(R)"             "PB(L5)"            "AOTU(R)"          
#>  [94] "PED(R)"            "g1(R)"             "NO(R)"            
#>  [97] "CRE(L)"            "SLP(R)"            "VLNP(R)"          
#> [100] "PRW"               "FLA(R)"            "NO1(R)"           
#> [103] "PB"                "SNP(R)"            "a'L(R)"           
#> [106] "BU(R)"             "ICL(L)"            "FBl8"             
#> [109] "SCL(L)"            "LAL(-GA)(R)"       "b2(R)"            
#> [112] "EBr1"              "VMNP"              "CA(R)"            
#> [115] "CAN(R)"            "LOP(R)"            "WED(R)"           
#> [118] "a2(R)"             "SIP(L)"            "b'L(R)"           
#> [121] "SCL(R)"            "PB(L1)"            "a'1(R)"           
#> [124] "SMP(R)"            "PLP(R)"            "INP"              
#> [127] "EPA(R)"            "gL(R)"             "FBl1"             
#> [130] "PB(R3)"            "FBl5"              "FBl6"             
#> [133] "g3(R)"             "EBr3am"            "mALT(L)"          
#> [136] "GOR(R)"            "AMMC"              "SPS(R)"           
#> [139] "BU(L)"             "PB(L8)"            "PB(L6)"           
#> [142] "FBl7"              "FB"                "LH(R)"            
#> [145] "PENP"              "SAD"               "PB(L7)"           
#> [148] "PB(R2)"            "NO(L)"             "a'3(R)"           
#> [151] "vACA(R)"          
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "g4(R)"             "NO"                "VLNP(R)"          
#>   [4] "AB(L)"             "AL-DA2(R)"         "AL-VP5(R)"        
#>   [7] "ATL(L)"            "PENP"              "a'L(R)"           
#>  [10] "ROB(R)"            "LX(L)"             "g3(R)"            
#>  [13] "NO1(R)"            "GF(R)"             "GA(R)"            
#>  [16] "AL-DL5(L)"         "EBr5"              "CRE(-RUB)(L)"     
#>  [19] "g5(R)"             "b'1(R)"            "FB"               
#>  [22] "AL-DM4(L)"         "NO(R)"             "AL-VA7l(R)"       
#>  [25] "LAL(R)"            "AL-DC4(R)"         "a1(R)"            
#>  [28] "bL(L)"             "PB(R3)"            "SMP(L)"           
#>  [31] "mALT(R)"           "AL-DM2(R)"         "AL-VM2(R)"        
#>  [34] "PB(L2)"            "BU(R)"             "AL-DC2(L)"        
#>  [37] "a'L(L)"            "AL-VP1l(R)"        "PB(L7)"           
#>  [40] "AL-VA4(R)"         "AL-VA5(R)"         "POC"              
#>  [43] "NO(L)"             "AL-VM5d(R)"        "PB(L5)"           
#>  [46] "EBr2r4"            "VES(R)"            "FBl1"             
#>  [49] "b'L(L)"            "AOTU(R)"           "CRE(L)"           
#>  [52] "AL-VM5v(R)"        "AMMC"              "dACA(R)"          
#>  [55] "ICL(R)"            "AL-DL3(R)"         "AL-VC4(R)"        
#>  [58] "g2(R)"             "AL-DM5(L)"         "PB(L9)"           
#>  [61] "PB(R2)"            "LAL(L)"            "LO(R)"            
#>  [64] "ME(R)"             "PB(R9)"            "PB(L3)"           
#>  [67] "AL-DM4(R)"         "AL-VA2(R)"         "EB"               
#>  [70] "AL-VM7d(L)"        "RUB(L)"            "AL-DP1l(R)"       
#>  [73] "AL-DP1m(R)"        "lACA(R)"           "FLA(R)"           
#>  [76] "FBl2"              "GNG"               "aL(R)"            
#>  [79] "AL-VA1v(R)"        "MB(L)"             "PB(L1)"           
#>  [82] "AL-DC1(R)"         "GC"                "a'1(R)"           
#>  [85] "FBl6"              "CA(R)"             "AL-DL2d(R)"       
#>  [88] "AL-VA7m(R)"        "LAL(-GA)(R)"       "EBr3pw"           
#>  [91] "AL-VM1(R)"         "CX"                "AL-DM2(L)"        
#>  [94] "FBl7"              "INP"               "AL-DM5(R)"        
#>  [97] "FBl5"              "AB(R)"             "CAN(R)"           
#> [100] "PB(R6)"            "SPS(R)"            "SNP(R)"           
#> [103] "CA(L)"             "PB(R8)"            "AL-VA6(L)"        
#> [106] "ATL(R)"            "RUB(R)"            "VES(L)"           
#> [109] "SLP(R)"            "AL-VM7v(R)"        "AL-VC1(R)"        
#> [112] "NO2(R)"            "a3(R)"             "AL-DA3(R)"        
#> [115] "a'2(R)"            "gL(L)"             "AL-DM6(L)"        
#> [118] "AL-VM7v(L)"        "AL-VM4(R)"         "PED(R)"           
#> [121] "AL-DL4(L)"         "FBl9"              "AL-DM6(R)"        
#> [124] "bL(R)"             "AL-VL1(R)"         "SCL(R)"           
#> [127] "AL-DM1(L)"         "IPS(R)"            "aL(L)"            
#> [130] "AL(L)"             "AL-DA3(L)"         "AL-VA6(R)"        
#> [133] "AL-VC2(R)"         "EPA(L)"            "AL-DA4m(R)"       
#> [136] "AL-VM7d(R)"        "PB(R5)"            "MB(+ACA)(R)"      
#> [139] "AL-VC5(R)"         "AL-VP1d(R)"        "GOR(L)"           
#> [142] "GOR(R)"            "AL-VC3l(R)"        "AL-D(R)"          
#> [145] "AL-VM3(R)"         "VMNP"              "AL-DC1(L)"        
#> [148] "SIP(R)"            "AL-DA1(R)"         "a2(R)"            
#> [151] "AL-DC3(R)"         "EBr3am"            "PB(R7)"           
#> [154] "SCL(L)"            "AL-DA2(L)"         "EBr6"             
#> [157] "AL-D(L)"           "AL-DL2v(R)"        "NO3(L)"           
#> [160] "PB(L6)"            "vACA(R)"           "AOT(R)"           
#> [163] "mALT(L)"           "NO3(R)"            "EBr1"             
#> [166] "PB(L8)"            "EPA(R)"            "MB(R)"            
#> [169] "PRW"               "SIP(L)"            "EBr3d"            
#> [172] "LH(R)"             "AL-DL5(R)"         "AL-VL2a(R)"       
#> [175] "SNP(L)"            "FBl8"              "AVLP(R)"          
#> [178] "b1(R)"             "PB(R1)"            "AME(R)"           
#> [181] "AL-VC3m(R)"        "AL-VP3(R)"         "SAD"              
#> [184] "g1(R)"             "NO2(L)"            "SMP(R)"           
#> [187] "a'3(R)"            "SPS(L)"            "AL-DC4(L)"        
#> [190] "AL-DP1m(L)"        "gL(R)"             "AL-VP1m(R)"       
#> [193] "CRE(-ROB,-RUB)(R)" "b2(R)"             "PVLP(R)"          
#> [196] "FB-column3"        "OL(R)"             "AL(R)"            
#> [199] "SAD(-AMMC)"        "AL-VA1d(R)"        "AL-VP2(R)"        
#> [202] "hemibrain"         "PB"                "AL-DA4l(R)"       
#> [205] "AL-VL2p(R)"        "b'2(R)"            "AL-DA4m(L)"       
#> [208] "FBl3"              "IB"                "PLP(R)"           
#> [211] "AL-DL4(R)"         "AL-VP4(R)"         "ICL(L)"           
#> [214] "FBl4"              "AL-DL1(R)"         "LOP(R)"           
#> [217] "BU(L)"             "AL-V(R)"           "AL-DM3(L)"        
#> [220] "AL-DM1(R)"         "AL-DM3(R)"         "AL-VA3(R)"        
#> [223] "PB(L4)"            "CRE(R)"            "b'L(R)"           
#> [226] "LX(R)"             "PB(R4)"            "WED(R)"           
#> [229] "AL-DC2(R)"         "NO1(L)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "PENP"              "MB(L)"             "EPA(R)"           
#>   [4] "VES(R)"            "LOP(R)"            "a'3(R)"           
#>   [7] "SIP(L)"            "a'L(R)"            "AL-VL2p(R)"       
#>  [10] "SCL(L)"            "PB(L5)"            "gL(L)"            
#>  [13] "SAD(-AMMC)"        "b'1(R)"            "NO(L)"            
#>  [16] "NO2(L)"            "AL-VC4(R)"         "PED(R)"           
#>  [19] "PB"                "AL-DA2(L)"         "NO2(R)"           
#>  [22] "GOR(R)"            "PLP(R)"            "AL-DL3(R)"        
#>  [25] "AL-DM1(R)"         "SAD"               "GNG"              
#>  [28] "VMNP"              "a'2(R)"            "PB(R9)"           
#>  [31] "OL(R)"             "AL-DL2d(R)"        "AL-VM2(R)"        
#>  [34] "SPS(R)"            "AL-DC2(L)"         "AL-DL5(L)"        
#>  [37] "PB(L6)"            "AL-VC5(R)"         "IB"               
#>  [40] "b'2(R)"            "GOR(L)"            "AL-VA2(R)"        
#>  [43] "AL-VC2(R)"         "g3(R)"             "a2(R)"            
#>  [46] "mALT(R)"           "AL-DA4m(L)"        "AL-VM1(R)"        
#>  [49] "ATL(L)"            "PB(L8)"            "SNP(R)"           
#>  [52] "EBr3pw"            "MB(+ACA)(R)"       "aL(L)"            
#>  [55] "AL-VA5(R)"         "ICL(L)"            "FBl1"             
#>  [58] "FBl9"              "AL-VM3(R)"         "AL-DM1(L)"        
#>  [61] "AL-DM4(L)"         "aL(R)"             "EBr3d"            
#>  [64] "FB-column3"        "ICL(R)"            "b'L(L)"           
#>  [67] "AL-VM7d(R)"        "VLNP(R)"           "AL-DC4(R)"        
#>  [70] "BU(R)"             "AL-DA4l(R)"        "AL-VA6(R)"        
#>  [73] "AL-VP3(R)"         "AL-VP1d(R)"        "dACA(R)"          
#>  [76] "SIP(R)"            "FBl3"              "AME(R)"           
#>  [79] "FLA(R)"            "AL-DM5(L)"         "FBl6"             
#>  [82] "LX(L)"             "PB(L4)"            "SMP(L)"           
#>  [85] "b1(R)"             "b'L(R)"            "EBr6"             
#>  [88] "SLP(R)"            "AL-DC3(R)"         "AL-DP1l(R)"       
#>  [91] "FBl7"              "LH(R)"             "AL-VC3m(R)"       
#>  [94] "SMP(R)"            "GF(R)"             "g1(R)"            
#>  [97] "AL-D(R)"           "PRW"               "EBr5"             
#> [100] "AL-VA4(R)"         "AL-VM7v(R)"        "PB(L9)"           
#> [103] "AOTU(R)"           "LX(R)"             "NO(R)"            
#> [106] "gL(R)"             "GA(R)"             "AL-V(R)"          
#> [109] "AB(L)"             "lACA(R)"           "PB(R8)"           
#> [112] "a'1(R)"            "VES(L)"            "b2(R)"            
#> [115] "BU(L)"             "AL-DC2(R)"         "AOT(R)"           
#> [118] "a1(R)"             "AL-VC3l(R)"        "vACA(R)"          
#> [121] "PB(R7)"            "EB"                "CRE(L)"           
#> [124] "AL-DM3(R)"         "AL-VA7l(R)"        "AL-VL1(R)"        
#> [127] "AL-VP1l(R)"        "CA(L)"             "SCL(R)"           
#> [130] "FBl2"              "MB(R)"             "PB(R6)"           
#> [133] "NO3(L)"            "PB(R3)"            "AL-DL5(R)"        
#> [136] "LAL(L)"            "EBr1"              "AL-DA3(L)"        
#> [139] "PB(R1)"            "PB(L7)"            "AL-DM6(R)"        
#> [142] "LAL(-GA)(R)"       "g2(R)"             "g4(R)"            
#> [145] "CA(R)"             "CAN(R)"            "PB(R4)"           
#> [148] "AL-DA1(R)"         "AL-DM2(L)"         "bL(R)"            
#> [151] "PB(R5)"            "NO3(R)"            "AL-VA1d(R)"       
#> [154] "AL-VP2(R)"         "AL-VM7d(L)"        "AL-DP1m(L)"       
#> [157] "AL-VP5(R)"         "EBr2r4"            "AL(L)"            
#> [160] "RUB(L)"            "AL-DL1(R)"         "NO"               
#> [163] "ATL(R)"            "AL-DM2(R)"         "mALT(L)"          
#> [166] "PVLP(R)"           "EPA(L)"            "AL-DM5(R)"        
#> [169] "AL-DL2v(R)"        "AL-VM5v(R)"        "AMMC"             
#> [172] "AL-D(L)"           "ROB(R)"            "AL(R)"            
#> [175] "AL-DC1(L)"         "AL-DL4(R)"         "hemibrain"        
#> [178] "SNP(L)"            "FBl4"              "PB(R2)"           
#> [181] "a'L(L)"            "AL-VA1v(R)"        "AL-VC1(R)"        
#> [184] "WED(R)"            "FB"                "AL-VA6(L)"        
#> [187] "a3(R)"             "AL-VP1m(R)"        "CX"               
#> [190] "CRE(R)"            "AL-DM3(L)"         "g5(R)"            
#> [193] "AVLP(R)"           "AL-DM4(R)"         "AL-VL2a(R)"       
#> [196] "EBr3am"            "AL-VP4(R)"         "PB(L3)"           
#> [199] "AL-DC4(L)"         "CRE(-RUB)(L)"      "AL-DA4m(R)"       
#> [202] "AL-DC1(R)"         "AL-VM4(R)"         "AL-VM5d(R)"       
#> [205] "POC"               "NO1(L)"            "SPS(L)"           
#> [208] "AL-DM6(L)"         "LO(R)"             "INP"              
#> [211] "AL-DA2(R)"         "AL-VA7m(R)"        "PB(L1)"           
#> [214] "bL(L)"             "CRE(-ROB,-RUB)(R)" "NO1(R)"           
#> [217] "IPS(R)"            "AL-DA3(R)"         "AL-DP1m(R)"       
#> [220] "PB(L2)"            "AL-VM7v(L)"        "FBl8"             
#> [223] "LAL(R)"            "RUB(R)"            "GC"               
#> [226] "FBl5"              "AB(R)"             "ME(R)"            
#> [229] "AL-DL4(L)"         "AL-VA3(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method     path                               
#> [1,] "POST"     "/api/v:ver/npexplorer/rankedtable"
#> [2,] "GET"      "/api/v:ver/cached/roiconnectivity"
#> [3,] "PATCH"    "/api"                             
#> [4,] "PROPFIND" "/api/*"                           
#> [5,] "GET"      "/*"                               
#> [6,] "HEAD"     "/api/*"                           
#>      name                                                                
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [3,] "github.com/labstack/echo.glob..func1"                              
#> [4,] "github.com/labstack/echo.glob..func1"                              
#> [5,] "github.com/labstack/echo.common.static.func1"                      
#> [6,] "github.com/labstack/echo.glob..func1"                              

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }