Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0`
#> $`hemibrain:v1.0`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0`$ROIs
#>   [1] "PB(R1)"            "FBl7"              "INP"              
#>   [4] "PB(R4)"            "CA(R)"             "RUB(L)"           
#>   [7] "b'1(R)"            "PRW"               "SIP(L)"           
#>  [10] "PED(R)"            "SPS(L)"            "EBr2r4"           
#>  [13] "a1(R)"             "b'2(R)"            "SMP(L)"           
#>  [16] "AOTU(R)"           "hemibrain"         "PB(L5)"           
#>  [19] "VLNP(R)"           "RUB(R)"            "NO1(L)"           
#>  [22] "a'1(R)"            "NO3(L)"            "EBr3pw"           
#>  [25] "OL(R)"             "MB(+ACA)(R)"       "IPS(R)"           
#>  [28] "NO(L)"             "PB(R5)"            "mALT(L)"          
#>  [31] "SPS(R)"            "EBr1"              "EBr3d"            
#>  [34] "gL(L)"             "PB(L6)"            "FBl1"             
#>  [37] "FBl3"              "AMMC"              "aL(R)"            
#>  [40] "gL(R)"             "FBl2"              "MB(R)"            
#>  [43] "CRE(-ROB,-RUB)(R)" "MB(L)"             "VMNP"             
#>  [46] "a'L(L)"            "NO3(R)"            "b2(R)"            
#>  [49] "AME(R)"            "POC"               "IB"               
#>  [52] "SNP(R)"            "SLP(R)"            "PB(R7)"           
#>  [55] "AVLP(R)"           "AOT(R)"            "WED(R)"           
#>  [58] "FB"                "g2(R)"             "PB(L7)"           
#>  [61] "PVLP(R)"           "b'L(L)"            "LOP(R)"           
#>  [64] "PB(R6)"            "LAL(-GA)(R)"       "CRE(R)"           
#>  [67] "PB(R9)"            "bL(R)"             "GA(R)"            
#>  [70] "BU(L)"             "bL(L)"             "FBl5"             
#>  [73] "g3(R)"             "dACA(R)"           "ICL(L)"           
#>  [76] "LAL(L)"            "PB(L3)"            "FBl8"             
#>  [79] "LAL(R)"            "SMP(R)"            "EPA(R)"           
#>  [82] "NO(R)"             "LX(L)"             "CAN(R)"           
#>  [85] "EBr3am"            "PB(L2)"            "SAD"              
#>  [88] "g1(R)"             "NO"                "PB(L9)"           
#>  [91] "AL(R)"             "SIP(R)"            "g5(R)"            
#>  [94] "VES(L)"            "a'3(R)"            "mALT(R)"          
#>  [97] "PB(L1)"            "PLP(R)"            "SNP(L)"           
#> [100] "LO(R)"             "ATL(L)"            "GOR(L)"           
#> [103] "EBr5"              "PB(R8)"            "GOR(R)"           
#> [106] "AB(L)"             "a2(R)"             "PB"               
#> [109] "LX(R)"             "vACA(R)"           "lACA(R)"          
#> [112] "ATL(R)"            "b1(R)"             "EPA(L)"           
#> [115] "PB(R2)"            "NO2(L)"            "b'L(R)"           
#> [118] "NO1(R)"            "EB"                "FBl9"             
#> [121] "LH(R)"             "FLA(R)"            "PB(L8)"           
#> [124] "AL(L)"             "ROB(R)"            "FBl6"             
#> [127] "PENP"              "VES(R)"            "AB(R)"            
#> [130] "PB(L4)"            "CX"                "SCL(L)"           
#> [133] "GC"                "FBl4"              "NO2(R)"           
#> [136] "aL(L)"             "a3(R)"             "ME(R)"            
#> [139] "PB(R3)"            "ICL(R)"            "CA(L)"            
#> [142] "GF(R)"             "a'2(R)"            "g4(R)"            
#> [145] "CRE(L)"            "SAD(-AMMC)"        "SCL(R)"           
#> [148] "EBr6"              "a'L(R)"            "BU(R)"            
#> [151] "GNG"              
#> 
#> $`hemibrain:v1.0`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                   
#> [1,] "GET"  "/oauth2callback"                      
#> [2,] "POST" "/api"                                 
#> [3,] "POST" "/api/v:ver/npexplorer/roiconnectivity"
#> [4,] "POST" "/api/v:ver/npexplorer/autapses"       
#> [5,] "POST" "/api/roimeshes/mesh/:dataset/:roi"    
#> [6,] "PUT"  "/api"                                 
#>      name                                                                    
#> [1,] "github.com/janelia-flyem/echo-secure.oauthCallbackHandler"             
#> [2,] "github.com/labstack/echo.glob..func1"                                  
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    
#> [5,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"
#> [6,] "github.com/labstack/echo.glob..func1"                                  

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }