Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AME(R)"            "AL-DC3(R)"         "VMNP"             
#>   [4] "CRE(-RUB)(L)"      "SCL(R)"            "PB"               
#>   [7] "AL-VC4(R)"         "SIP(R)"            "AB(R)"            
#>  [10] "g5(R)"             "a'3(R)"            "FBl1"             
#>  [13] "AL-VM4(R)"         "AL-DP1m(R)"        "a2(R)"            
#>  [16] "AL-DA2(R)"         "ATL(L)"            "AL-DC1(R)"        
#>  [19] "AL-DL2v(R)"        "MB(R)"             "AL-VP1d(R)"       
#>  [22] "ICL(R)"            "a'L(L)"            "AL-VA3(R)"        
#>  [25] "VLNP(R)"           "EBr1"              "EBr3d"            
#>  [28] "AL-DA3(R)"         "EBr2r4"            "AL-DC4(L)"        
#>  [31] "GNG"               "vACA(R)"           "AL-DM2(R)"        
#>  [34] "AL-DM4(R)"         "AL-VA1v(R)"        "AL-VA5(R)"        
#>  [37] "LX(L)"             "AL-DL5(R)"         "AL-DM6(L)"        
#>  [40] "PB(L4)"            "AL-V(R)"           "BU(R)"            
#>  [43] "RUB(L)"            "FB-column3"        "PB(L5)"           
#>  [46] "ME(R)"             "EPA(R)"            "FBl6"             
#>  [49] "AL-VM1(R)"         "LAL(-GA)(R)"       "FBl4"             
#>  [52] "NO2(R)"            "SCL(L)"            "aL(R)"            
#>  [55] "AL(L)"             "AL-VM5v(R)"        "AL-VM7d(R)"       
#>  [58] "POC"               "AL-DL4(L)"         "PB(R9)"           
#>  [61] "AL-D(R)"           "SMP(L)"            "AL-VM5d(R)"       
#>  [64] "b'1(R)"            "CX"                "CA(L)"            
#>  [67] "PRW"               "b'L(R)"            "AL-DM6(R)"        
#>  [70] "AL-VP5(R)"         "AL-VP1l(R)"        "PENP"             
#>  [73] "AL(R)"             "AL-DL2d(R)"        "FLA(R)"           
#>  [76] "AL-D(L)"           "gL(R)"             "AL-VC5(R)"        
#>  [79] "PB(L3)"            "SPS(L)"            "AB(L)"            
#>  [82] "EBr3pw"            "AL-VL2p(R)"        "bL(R)"            
#>  [85] "RUB(R)"            "LO(R)"             "FBl8"             
#>  [88] "MB(+ACA)(R)"       "AL-VA2(R)"         "SPS(R)"           
#>  [91] "LX(R)"             "NO(R)"             "PB(L7)"           
#>  [94] "LAL(L)"            "PB(L6)"            "SNP(L)"           
#>  [97] "PVLP(R)"           "AL-VL2a(R)"        "g3(R)"            
#> [100] "AL-DC1(L)"         "EB"                "AVLP(R)"          
#> [103] "EPA(L)"            "GOR(R)"            "dACA(R)"          
#> [106] "PLP(R)"            "NO1(R)"            "AL-DA4m(L)"       
#> [109] "AL-DP1m(L)"        "AOT(R)"            "PB(L2)"           
#> [112] "SIP(L)"            "GC"                "AL-VM3(R)"        
#> [115] "AL-VM7v(L)"        "AL-VA7l(R)"        "PB(L9)"           
#> [118] "AL-DA4m(R)"        "AL-DP1l(R)"        "EBr5"             
#> [121] "AL-DL5(L)"         "AL-DM5(R)"         "AL-VC1(R)"        
#> [124] "AL-VC3m(R)"        "VES(R)"            "FBl3"             
#> [127] "AL-VA7m(R)"        "AL-DM1(L)"         "SAD(-AMMC)"       
#> [130] "aL(L)"             "AL-VA1d(R)"        "a1(R)"            
#> [133] "SAD"               "ATL(R)"            "AL-VA4(R)"        
#> [136] "AL-VP4(R)"         "g1(R)"             "FB"               
#> [139] "AL-DM5(L)"         "PB(R1)"            "EBr6"             
#> [142] "SNP(R)"            "IB112"             "b1(R)"            
#> [145] "AL-DA4l(R)"        "WED(R)"            "CRE(L)"           
#> [148] "PB(R3)"            "a3(R)"             "AL-VC2(R)"        
#> [151] "SMP(R)"            "AMMC"              "AL-VA6(L)"        
#> [154] "NO"                "b'2(R)"            "AL-DL3(R)"        
#> [157] "AL-VM7d(L)"        "b'L(L)"            "gL(L)"            
#> [160] "AL-DM2(L)"         "LH(R)"             "PB(R2)"           
#> [163] "SLP(R)"            "b2(R)"             "bL(L)"            
#> [166] "CA(R)"             "AL-VM7v(R)"        "PB(L1)"           
#> [169] "LAL(R)"            "mALT(L)"           "NO2(L)"           
#> [172] "AL-DM3(R)"         "GOR(L)"            "AL-DM1(R)"        
#> [175] "FBl9"              "ROB(R)"            "a'1(R)"           
#> [178] "a'2(R)"            "FBl5"              "OL(R)"            
#> [181] "AL-DA1(R)"         "AL-VL1(R)"         "PB(R6)"           
#> [184] "lACA(R)"           "IB"                "AL-DA3(L)"        
#> [187] "AL-DC2(R)"         "ICL(L)"            "FBl7"             
#> [190] "PB(R4)"            "MB(L)"             "BU(L)"            
#> [193] "CRE(R)"            "AL-DC4(R)"         "CRE(-ROB,-RUB)(R)"
#> [196] "GF(R)"             "GA(R)"             "AL-VA6(R)"        
#> [199] "AL-VP3(R)"         "g2(R)"             "AL-DM3(L)"        
#> [202] "EBr3am"            "PED(R)"            "AL-DA2(L)"        
#> [205] "PB(R5)"            "NO(L)"             "VES(L)"           
#> [208] "PB(L8)"            "CAN(R)"            "AL-VM2(R)"        
#> [211] "AOTU(R)"           "NO3(R)"            "AL-DC2(L)"        
#> [214] "LOP(R)"            "g4(R)"             "AL-DL4(R)"        
#> [217] "mALT(R)"           "PB(R8)"            "FBl2"             
#> [220] "IPS(R)"            "INP"               "AL-VP2(R)"        
#> [223] "hemibrain"         "NO3(L)"            "AL-VP1m(R)"       
#> [226] "PB(R7)"            "AL-DL1(R)"         "AL-DM4(L)"        
#> [229] "AL-VC3l(R)"        "NO1(L)"            "a'L(R)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "FBl6"              "LAL(L)"            "GOR(L)"           
#>   [4] "LX(R)"             "AB(L)"             "ME(R)"            
#>   [7] "a'L(R)"            "ICL(R)"            "LOP(R)"           
#>  [10] "a1(R)"             "PB(R8)"            "EBr3pw"           
#>  [13] "NO(L)"             "AME(R)"            "RUB(L)"           
#>  [16] "IB"                "mALT(L)"           "AOTU(R)"          
#>  [19] "EBr1"              "NO3(L)"            "CA(L)"            
#>  [22] "PB(L8)"            "FBl2"              "PB(L5)"           
#>  [25] "LO(R)"             "VES(R)"            "FBl7"             
#>  [28] "PB(L3)"            "SCL(R)"            "PB"               
#>  [31] "SMP(L)"            "CRE(R)"            "FBl4"             
#>  [34] "b2(R)"             "NO"                "g3(R)"            
#>  [37] "PRW"               "LAL(-GA)(R)"       "PED(R)"           
#>  [40] "PB(L6)"            "VLNP(R)"           "MB(L)"            
#>  [43] "EBr5"              "VMNP"              "MB(+ACA)(R)"      
#>  [46] "ATL(R)"            "GF(R)"             "SAD"              
#>  [49] "SIP(L)"            "CRE(-ROB,-RUB)(R)" "SAD(-AMMC)"       
#>  [52] "FBl3"              "b'1(R)"            "SMP(R)"           
#>  [55] "b'L(R)"            "BU(L)"             "GNG"              
#>  [58] "EPA(R)"            "NO(R)"             "b'2(R)"           
#>  [61] "PB(R7)"            "PB(L1)"            "FB"               
#>  [64] "bL(L)"             "RUB(R)"            "FBl5"             
#>  [67] "GOR(R)"            "PB(R3)"            "VES(L)"           
#>  [70] "NO1(R)"            "EBr2r4"            "FBl9"             
#>  [73] "EPA(L)"            "MB(R)"             "AOT(R)"           
#>  [76] "a'3(R)"            "PB(R6)"            "g4(R)"            
#>  [79] "PB(L7)"            "PB(R1)"            "CA(R)"            
#>  [82] "a'L(L)"            "dACA(R)"           "EBr3d"            
#>  [85] "ROB(R)"            "gL(R)"             "AVLP(R)"          
#>  [88] "IPS(R)"            "EBr3am"            "aL(R)"            
#>  [91] "EBr6"              "SNP(R)"            "g2(R)"            
#>  [94] "SNP(L)"            "b'L(L)"            "PLP(R)"           
#>  [97] "ICL(L)"            "PVLP(R)"           "AB(R)"            
#> [100] "NO2(R)"            "a2(R)"             "mALT(R)"          
#> [103] "GC"                "ATL(L)"            "b1(R)"            
#> [106] "bL(R)"             "SLP(R)"            "AL(R)"            
#> [109] "hemibrain"         "gL(L)"             "g5(R)"            
#> [112] "aL(L)"             "LAL(R)"            "lACA(R)"          
#> [115] "BU(R)"             "PB(R2)"            "NO1(L)"           
#> [118] "AL(L)"             "LX(L)"             "CAN(R)"           
#> [121] "SIP(R)"            "PB(L9)"            "FBl8"             
#> [124] "LH(R)"             "POC"               "PENP"             
#> [127] "WED(R)"            "NO3(R)"            "a'1(R)"           
#> [130] "FBl1"              "INP"               "NO2(L)"           
#> [133] "g1(R)"             "FLA(R)"            "SPS(L)"           
#> [136] "PB(R9)"            "a3(R)"             "OL(R)"            
#> [139] "PB(L2)"            "vACA(R)"           "AMMC"             
#> [142] "EB"                "CRE(L)"            "PB(R4)"           
#> [145] "PB(R5)"            "SCL(L)"            "GA(R)"            
#> [148] "SPS(R)"            "a'2(R)"            "PB(L4)"           
#> [151] "CX"               
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "FB-column3"        "AL-DL2v(R)"        "POC"              
#>   [4] "EBr1"              "CRE(L)"            "MB(+ACA)(R)"      
#>   [7] "PED(R)"            "AL-DP1m(L)"        "IPS(R)"           
#>  [10] "b'2(R)"            "AL(L)"             "SIP(R)"           
#>  [13] "AL-DA3(L)"         "NO1(L)"            "NO1(R)"           
#>  [16] "SNP(R)"            "EBr5"              "a'2(R)"           
#>  [19] "AL-DA3(R)"         "ATL(R)"            "AL-VA4(R)"        
#>  [22] "AL-VM7v(R)"        "PB(L5)"            "GOR(L)"           
#>  [25] "AL-DA2(L)"         "PB(L9)"            "AL-DM3(L)"        
#>  [28] "SLP(R)"            "AL-VM7d(R)"        "SMP(L)"           
#>  [31] "NO3(R)"            "g4(R)"             "AL-DL5(R)"        
#>  [34] "PB(L8)"            "FBl1"              "GOR(R)"           
#>  [37] "PB(R3)"            "AL-DP1m(R)"        "AL-VP1m(R)"       
#>  [40] "NO(L)"             "AL-VC3m(R)"        "BU(R)"            
#>  [43] "AL-DM2(L)"         "AL-DM4(L)"         "LH(R)"            
#>  [46] "a'L(L)"            "AL-DL1(R)"         "PB(R7)"           
#>  [49] "AL-DM1(L)"         "ROB(R)"            "SPS(L)"           
#>  [52] "g1(R)"             "b'1(R)"            "AL-DM5(L)"        
#>  [55] "AL-VM7d(L)"        "LO(R)"             "LX(R)"            
#>  [58] "AL-DC2(L)"         "RUB(R)"            "CA(L)"            
#>  [61] "PRW"               "AL-VP5(R)"         "AL-VP3(R)"        
#>  [64] "AL-VC5(R)"         "SNP(L)"            "AL-DM3(R)"        
#>  [67] "AL-VA2(R)"         "CRE(-RUB)(L)"      "LX(L)"            
#>  [70] "EPA(L)"            "AL-DM4(R)"         "AL-VC3l(R)"       
#>  [73] "GF(R)"             "SIP(L)"            "EB"               
#>  [76] "EBr3am"            "LAL(-GA)(R)"       "GC"               
#>  [79] "AL-VC1(R)"         "AL-DC2(R)"         "NO(R)"            
#>  [82] "FBl9"              "AL-VM7v(L)"        "AB(R)"            
#>  [85] "AL-VP1l(R)"        "AL-DA4m(R)"        "AL-VA3(R)"        
#>  [88] "vACA(R)"           "AL-DC1(L)"         "SAD(-AMMC)"       
#>  [91] "PVLP(R)"           "CX"                "FLA(R)"           
#>  [94] "AVLP(R)"           "AL-VA6(R)"         "AL-DM6(R)"        
#>  [97] "AL-D(L)"           "AL-VP4(R)"         "AL-VP2(R)"        
#> [100] "gL(L)"             "FBl8"              "a2(R)"            
#> [103] "GA(R)"             "EBr3d"             "AL-VM1(R)"        
#> [106] "CRE(-ROB,-RUB)(R)" "a'3(R)"            "VES(L)"           
#> [109] "AL-DC3(R)"         "AL-DM5(R)"         "PB(R1)"           
#> [112] "INP"               "RUB(L)"            "NO3(L)"           
#> [115] "GNG"               "AL-VM5v(R)"        "AL-VA5(R)"        
#> [118] "AL-VM2(R)"         "NO"                "FBl6"             
#> [121] "AL-DL4(R)"         "WED(R)"            "AL-VM4(R)"        
#> [124] "gL(R)"             "ICL(R)"            "OL(R)"            
#> [127] "aL(R)"             "BU(L)"             "a'L(R)"           
#> [130] "a3(R)"             "AL-DC1(R)"         "AL-VL1(R)"        
#> [133] "AL-VL2a(R)"        "hemibrain"         "PB(L1)"           
#> [136] "AL-DM6(L)"         "AL-VM5d(R)"        "AL(R)"            
#> [139] "AL-DM2(R)"         "AL-VC2(R)"         "PB(R5)"           
#> [142] "ATL(L)"            "LAL(L)"            "AL-VA6(L)"        
#> [145] "aL(L)"             "FBl2"              "CAN(R)"           
#> [148] "LAL(R)"            "dACA(R)"           "bL(L)"            
#> [151] "VES(R)"            "ME(R)"             "FB"               
#> [154] "VMNP"              "EBr2r4"            "AL-DL4(L)"        
#> [157] "PB(R4)"            "NO2(L)"            "AL-DA1(R)"        
#> [160] "AL-VA1d(R)"        "PB(L3)"            "AOTU(R)"          
#> [163] "AL-VM3(R)"         "SCL(R)"            "EPA(R)"           
#> [166] "AL-VA7l(R)"        "FBl7"              "AL-DM1(R)"        
#> [169] "AL-VC4(R)"         "a1(R)"             "EBr6"             
#> [172] "AL-DL5(L)"         "ICL(L)"            "AMMC"             
#> [175] "PB"                "FBl5"              "AL-V(R)"          
#> [178] "AL-DC4(R)"         "SPS(R)"            "PB(R8)"           
#> [181] "g5(R)"             "b'L(R)"            "AL-DA4m(L)"       
#> [184] "AL-DP1l(R)"        "b2(R)"             "AL-DL3(R)"        
#> [187] "AL-VA1v(R)"        "AL-VL2p(R)"        "PB(L7)"           
#> [190] "CA(R)"             "b'L(L)"            "mALT(R)"          
#> [193] "PB(L6)"            "b1(R)"             "PB(L4)"           
#> [196] "MB(R)"             "PENP"              "PB(R6)"           
#> [199] "g3(R)"             "AB(L)"             "AL-VP1d(R)"       
#> [202] "IB"                "lACA(R)"           "CRE(R)"           
#> [205] "NO2(R)"            "bL(R)"             "AOT(R)"           
#> [208] "PB(L2)"            "PLP(R)"            "AL-DC4(L)"        
#> [211] "SCL(L)"            "AL-DA2(R)"         "AL-D(R)"          
#> [214] "VLNP(R)"           "LOP(R)"            "a'1(R)"           
#> [217] "FBl3"              "AL-VA7m(R)"        "MB(L)"            
#> [220] "AL-DA4l(R)"        "EBr3pw"            "SMP(R)"           
#> [223] "g2(R)"             "PB(R2)"            "SAD"              
#> [226] "AL-DL2d(R)"        "mALT(L)"           "PB(R9)"           
#> [229] "FBl4"              "AME(R)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-VL2p(R)"        "a'L(R)"            "RUB(L)"           
#>   [4] "AL-DM3(R)"         "AL-VM5d(R)"        "a'L(L)"           
#>   [7] "AL-VC4(R)"         "AMMC"              "b'L(R)"           
#>  [10] "bL(R)"             "AL(R)"             "PB(L8)"           
#>  [13] "AL-DA2(L)"         "AL-DM1(L)"         "AL-DA3(L)"        
#>  [16] "PB(R1)"            "FBl8"              "CA(R)"            
#>  [19] "a'2(R)"            "AL-DM6(L)"         "FBl6"             
#>  [22] "LX(R)"             "AL-DP1m(L)"        "CRE(R)"           
#>  [25] "NO3(L)"            "FBl5"              "AL-DC4(L)"        
#>  [28] "EPA(L)"            "AL-DC2(L)"         "AL-DM3(L)"        
#>  [31] "POC"               "PB(R3)"            "OL(R)"            
#>  [34] "AL-DC3(R)"         "LO(R)"             "VES(R)"           
#>  [37] "FBl4"              "MB(R)"             "AL-VM1(R)"        
#>  [40] "a'3(R)"            "AL-DL4(L)"         "ICL(L)"           
#>  [43] "PB(R4)"            "AL-DM6(R)"         "AL-VA7m(R)"       
#>  [46] "RUB(R)"            "AL-VC3m(R)"        "AL-VM7d(R)"       
#>  [49] "AL-VP2(R)"         "AOT(R)"            "LAL(L)"           
#>  [52] "PB(L3)"            "AL-DM1(R)"         "AL-D(R)"          
#>  [55] "GA(R)"             "AL-DM2(L)"         "EBr3pw"           
#>  [58] "EBr6"              "a1(R)"             "g1(R)"            
#>  [61] "aL(L)"             "mALT(L)"           "EPA(R)"           
#>  [64] "AL-DA4l(R)"        "GC"                "BU(L)"            
#>  [67] "PB(L1)"            "AL-VM2(R)"         "AL-DC1(L)"        
#>  [70] "LH(R)"             "b'L(L)"            "AL-DC1(R)"        
#>  [73] "AL-VM7v(R)"        "SIP(L)"            "PB(R8)"           
#>  [76] "BU(R)"             "SMP(R)"            "ROB(R)"           
#>  [79] "a'1(R)"            "lACA(R)"           "g3(R)"            
#>  [82] "FBl9"              "AVLP(R)"           "AL-DC4(R)"        
#>  [85] "FB"                "PB(R5)"            "SAD(-AMMC)"       
#>  [88] "FBl2"              "AL-VL1(R)"         "AL-VP5(R)"        
#>  [91] "vACA(R)"           "GOR(L)"            "gL(L)"            
#>  [94] "AL-DA4m(L)"        "AL-DA3(R)"         "AL-VA4(R)"        
#>  [97] "NO"                "bL(L)"             "FB-column3"       
#> [100] "AL-VM7v(L)"        "gL(R)"             "AL-DL3(R)"        
#> [103] "SCL(L)"            "dACA(R)"           "EBr3d"            
#> [106] "INP"               "CX"                "MB(+ACA)(R)"      
#> [109] "CAN(R)"            "AL-VM5v(R)"        "AL-DM4(L)"        
#> [112] "AL-VP3(R)"         "PB(L2)"            "AL-VA5(R)"        
#> [115] "AL-VC1(R)"         "AL-VM3(R)"         "g5(R)"            
#> [118] "ATL(R)"            "NO(L)"             "g4(R)"            
#> [121] "AL-DL1(R)"         "VLNP(R)"           "mALT(R)"          
#> [124] "AL-DA4m(R)"        "CA(L)"             "GF(R)"            
#> [127] "LOP(R)"            "NO2(R)"            "AL-DM4(R)"        
#> [130] "AL-VA2(R)"         "EBr3am"            "PB(R6)"           
#> [133] "AL-DP1m(R)"        "AL-VC5(R)"         "PVLP(R)"          
#> [136] "AL-DM2(R)"         "CRE(-ROB,-RUB)(R)" "a2(R)"            
#> [139] "EBr2r4"            "NO(R)"             "FBl3"             
#> [142] "hemibrain"         "PED(R)"            "VES(L)"           
#> [145] "AL-DM5(R)"         "PB(R9)"            "FBl7"             
#> [148] "PB(L5)"            "NO1(L)"            "SCL(R)"           
#> [151] "AL-DP1l(R)"        "PB(L6)"            "CRE(L)"           
#> [154] "PB(L7)"            "SNP(L)"            "GNG"              
#> [157] "SPS(R)"            "LAL(-GA)(R)"       "EBr5"             
#> [160] "LAL(R)"            "AL-VA1v(R)"        "ATL(L)"           
#> [163] "AOTU(R)"           "PB(R2)"            "b'1(R)"           
#> [166] "LX(L)"             "AL-VA3(R)"         "AL-VA7l(R)"       
#> [169] "AL-VL2a(R)"        "PRW"               "VMNP"             
#> [172] "ME(R)"             "AL-DL2v(R)"        "AB(L)"            
#> [175] "PENP"              "AL-D(L)"           "AL-DM5(L)"        
#> [178] "b'2(R)"            "SMP(L)"            "AL-DL2d(R)"       
#> [181] "SNP(R)"            "CRE(-RUB)(L)"      "IB"               
#> [184] "NO2(L)"            "g2(R)"             "PB(L4)"           
#> [187] "AL-VP4(R)"         "EB"                "PB(L9)"           
#> [190] "ICL(R)"            "AL-DA2(R)"         "AL-VM4(R)"        
#> [193] "aL(R)"             "b2(R)"             "NO1(R)"           
#> [196] "AL-V(R)"           "AL-VP1m(R)"        "AL-VP1l(R)"       
#> [199] "AB(R)"             "SLP(R)"            "AL-DL4(R)"        
#> [202] "FLA(R)"            "PB"                "AL-DC2(R)"        
#> [205] "SPS(L)"            "EBr1"              "GOR(R)"           
#> [208] "AL-DL5(R)"         "SAD"               "WED(R)"           
#> [211] "NO3(R)"            "FBl1"              "PB(R7)"           
#> [214] "MB(L)"             "AL-DL5(L)"         "a3(R)"            
#> [217] "AL-VA1d(R)"        "AL-VA6(L)"         "SIP(R)"           
#> [220] "b1(R)"             "PLP(R)"            "AL-DA1(R)"        
#> [223] "AL-VA6(R)"         "AL(L)"             "AL-VC3l(R)"       
#> [226] "AL-VC2(R)"         "AL-VP1d(R)"        "AL-VM7d(L)"       
#> [229] "IPS(R)"            "AME(R)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method    path                                      
#> [1,] "POST"    "/logout"                                 
#> [2,] "CONNECT" "/api/*"                                  
#> [3,] "POST"    "/api/v:ver/npexplorer/roiconnectivity"   
#> [4,] "GET"     "/api/available"                          
#> [5,] "GET"     "/api/v:ver/dbmeta/database"              
#> [6,] "POST"    "/api/v:ver/npexplorer/commonconnectivity"
#>      name                                                                    
#> [1,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"
#> [2,] "github.com/labstack/echo.glob..func1"                                  
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.SetupRoutes.func1"     
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }