Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "g5(R)"             "AL-VC2(R)"         "AL-VM1(R)"        
#>   [4] "AOT(R)"            "EBr1"              "AL-DM3(L)"        
#>   [7] "FBl8"              "AL-DM1(R)"         "AL-VC3l(R)"       
#>  [10] "NO3(R)"            "g2(R)"             "AL-DC1(L)"        
#>  [13] "AL-DC4(R)"         "NO2(L)"            "AME(R)"           
#>  [16] "AL-VL1(R)"         "ATL(L)"            "SAD"              
#>  [19] "GOR(L)"            "AL(L)"             "AL-DM5(L)"        
#>  [22] "NO(R)"             "FBl6"              "ATL(R)"           
#>  [25] "ME(R)"             "NO3(L)"            "RUB(L)"           
#>  [28] "POC"               "SMP(L)"            "SIP(L)"           
#>  [31] "IB112"             "CAN(R)"            "AL-DC4(L)"        
#>  [34] "SLP(R)"            "AL-DA3(R)"         "AL-VM5d(R)"       
#>  [37] "PB(L3)"            "PB(R9)"            "CRE(-RUB)(L)"     
#>  [40] "VLNP(R)"           "AL-VA2(R)"         "GA(R)"            
#>  [43] "AL-DA1(R)"         "AL-DM3(R)"         "SMP(R)"           
#>  [46] "PB(L2)"            "AL-VA6(R)"         "AL-VP5(R)"        
#>  [49] "CRE(-ROB,-RUB)(R)" "AL-DA2(L)"         "FB-column3"       
#>  [52] "PB(L4)"            "AL-DM4(R)"         "vACA(R)"          
#>  [55] "FLA(R)"            "AL-DL4(L)"         "GF(R)"            
#>  [58] "NO2(R)"            "b2(R)"             "bL(L)"            
#>  [61] "FBl7"              "PB(R2)"            "AL-VA3(R)"        
#>  [64] "FBl2"              "AL-VA1d(R)"        "LOP(R)"           
#>  [67] "PENP"              "AOTU(R)"           "SNP(R)"           
#>  [70] "AL-VC5(R)"         "EBr5"              "bL(R)"            
#>  [73] "AL-DL2v(R)"        "PB(R5)"            "b'L(R)"           
#>  [76] "FBl9"              "AL-DL5(R)"         "BU(R)"            
#>  [79] "CRE(L)"            "AVLP(R)"           "GNG"              
#>  [82] "AL-DP1l(R)"        "SPS(R)"            "PB(R8)"           
#>  [85] "a3(R)"             "NO(L)"             "AL-DM5(R)"        
#>  [88] "ICL(R)"            "PB(L6)"            "PVLP(R)"          
#>  [91] "AL-DL5(L)"         "SIP(R)"            "AL(R)"            
#>  [94] "AL-VL2p(R)"        "AL-VP2(R)"         "SCL(R)"           
#>  [97] "AL-DM4(L)"         "VES(L)"            "MB(R)"            
#> [100] "AL-DC2(L)"         "AL-VM7v(L)"        "PB(R6)"           
#> [103] "AL-DM1(L)"         "gL(R)"             "g1(R)"            
#> [106] "LX(R)"             "AL-DC2(R)"         "AL-VA5(R)"        
#> [109] "CX"                "EBr6"              "AL-DA4m(R)"       
#> [112] "SCL(L)"            "b'2(R)"            "PB(L8)"           
#> [115] "LH(R)"             "AL-VP1m(R)"        "AL-VA1v(R)"       
#> [118] "WED(R)"            "AL-VM7d(L)"        "AL-VA7l(R)"       
#> [121] "a'3(R)"            "a'1(R)"            "b'L(L)"           
#> [124] "AL-DL2d(R)"        "AL-VA7m(R)"        "NO1(L)"           
#> [127] "EBr3d"             "ICL(L)"            "FBl3"             
#> [130] "OL(R)"             "AL-VP1d(R)"        "AL-VM2(R)"        
#> [133] "SAD(-AMMC)"        "AL-VP1l(R)"        "GC"               
#> [136] "AL-VP3(R)"         "lACA(R)"           "FB"               
#> [139] "PB"                "ROB(R)"            "AL-VM7v(R)"       
#> [142] "AL-V(R)"           "AL-DM2(L)"         "FBl5"             
#> [145] "AL-VL2a(R)"        "PB(L5)"            "CRE(R)"           
#> [148] "AL-DA4m(L)"        "AB(R)"             "AL-D(R)"          
#> [151] "AL-VM4(R)"         "MB(L)"             "PB(R4)"           
#> [154] "AL-DL1(R)"         "AL-VP4(R)"         "CA(L)"            
#> [157] "b1(R)"             "AL-VM5v(R)"        "b'1(R)"           
#> [160] "EBr3am"            "g4(R)"             "INP"              
#> [163] "AL-DM6(R)"         "PB(L9)"            "VMNP"             
#> [166] "EBr2r4"            "PB(R3)"            "AL-VC3m(R)"       
#> [169] "a2(R)"             "EB"                "a'L(R)"           
#> [172] "AL-DA3(L)"         "IPS(R)"            "IB"               
#> [175] "aL(R)"             "PRW"               "SNP(L)"           
#> [178] "AL-DL3(R)"         "AL-VC1(R)"         "AL-VM7d(R)"       
#> [181] "AB(L)"             "hemibrain"         "gL(L)"            
#> [184] "AL-D(L)"           "AL-VA6(L)"         "mALT(L)"          
#> [187] "LAL(-GA)(R)"       "LO(R)"             "PB(L7)"           
#> [190] "AL-DM2(R)"         "AL-DP1m(R)"        "AL-VA4(R)"        
#> [193] "EPA(R)"            "AL-DP1m(L)"        "VES(R)"           
#> [196] "aL(L)"             "AL-DC1(R)"         "PB(R7)"           
#> [199] "AL-DM6(L)"         "CA(R)"             "AL-DL4(R)"        
#> [202] "a1(R)"             "PLP(R)"            "MB(+ACA)(R)"      
#> [205] "AL-VC4(R)"         "mALT(R)"           "BU(L)"            
#> [208] "LAL(R)"            "RUB(R)"            "LAL(L)"           
#> [211] "FBl4"              "AL-VM3(R)"         "g3(R)"            
#> [214] "SPS(L)"            "AL-DC3(R)"         "dACA(R)"          
#> [217] "NO1(R)"            "FBl1"              "EPA(L)"           
#> [220] "a'L(L)"            "AL-DA2(R)"         "PB(R1)"           
#> [223] "LX(L)"             "PED(R)"            "PB(L1)"           
#> [226] "NO"                "EBr3pw"            "GOR(R)"           
#> [229] "AL-DA4l(R)"        "AMMC"              "a'2(R)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "mALT(R)"           "BU(L)"             "FBl1"             
#>   [4] "b'1(R)"            "PED(R)"            "FB"               
#>   [7] "LO(R)"             "GOR(L)"            "FBl6"             
#>  [10] "AME(R)"            "a3(R)"             "OL(R)"            
#>  [13] "MB(R)"             "b2(R)"             "PB(L1)"           
#>  [16] "EPA(L)"            "PRW"               "SCL(R)"           
#>  [19] "PB(R9)"            "g2(R)"             "PB(R8)"           
#>  [22] "a'L(L)"            "PB(L2)"            "PB(R2)"           
#>  [25] "GC"                "NO1(R)"            "FBl4"             
#>  [28] "PB(L4)"            "AOT(R)"            "PB(R4)"           
#>  [31] "aL(R)"             "NO3(R)"            "LX(R)"            
#>  [34] "SNP(R)"            "SIP(R)"            "FBl5"             
#>  [37] "CA(R)"             "EPA(R)"            "GOR(R)"           
#>  [40] "AB(L)"             "LAL(R)"            "lACA(R)"          
#>  [43] "g1(R)"             "PB(L8)"            "PB"               
#>  [46] "g4(R)"             "CRE(L)"            "a'1(R)"           
#>  [49] "AL(R)"             "CRE(-ROB,-RUB)(R)" "SAD"              
#>  [52] "EBr2r4"            "g5(R)"             "ATL(R)"           
#>  [55] "b1(R)"             "IB"                "g3(R)"            
#>  [58] "SPS(R)"            "gL(R)"             "AVLP(R)"          
#>  [61] "ICL(R)"            "LX(L)"             "a'2(R)"           
#>  [64] "gL(L)"             "NO"                "EBr3pw"           
#>  [67] "VES(R)"            "FBl9"              "AB(R)"            
#>  [70] "PB(R5)"            "b'2(R)"            "CAN(R)"           
#>  [73] "VLNP(R)"           "MB(L)"             "NO(R)"            
#>  [76] "SLP(R)"            "ATL(L)"            "PB(R7)"           
#>  [79] "VMNP"              "b'L(L)"            "a'3(R)"           
#>  [82] "AMMC"              "AOTU(R)"           "BU(R)"            
#>  [85] "PB(L7)"            "SNP(L)"            "NO(L)"            
#>  [88] "NO2(L)"            "PB(R6)"            "NO3(L)"           
#>  [91] "IPS(R)"            "aL(L)"             "INP"              
#>  [94] "LAL(L)"            "hemibrain"         "ME(R)"            
#>  [97] "ROB(R)"            "PB(L6)"            "bL(L)"            
#> [100] "LH(R)"             "LAL(-GA)(R)"       "CRE(R)"           
#> [103] "ICL(L)"            "PB(R3)"            "vACA(R)"          
#> [106] "a1(R)"             "dACA(R)"           "PVLP(R)"          
#> [109] "FBl2"              "FBl7"              "RUB(L)"           
#> [112] "POC"               "RUB(R)"            "NO1(L)"           
#> [115] "PB(L9)"            "SAD(-AMMC)"        "SMP(R)"           
#> [118] "SCL(L)"            "VES(L)"            "SIP(L)"           
#> [121] "PLP(R)"            "EB"                "PB(R1)"           
#> [124] "FBl8"              "LOP(R)"            "CA(L)"            
#> [127] "PENP"              "GF(R)"             "EBr5"             
#> [130] "SPS(L)"            "NO2(R)"            "EBr3am"           
#> [133] "a2(R)"             "EBr3d"             "MB(+ACA)(R)"      
#> [136] "GA(R)"             "EBr1"              "EBr6"             
#> [139] "FLA(R)"            "FBl3"              "PB(L3)"           
#> [142] "b'L(R)"            "GNG"               "a'L(R)"           
#> [145] "SMP(L)"            "AL(L)"             "bL(R)"            
#> [148] "CX"                "WED(R)"            "PB(L5)"           
#> [151] "mALT(L)"          
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "NO(L)"             "AL-DA2(R)"         "AL-VP5(R)"        
#>   [4] "mALT(L)"           "NO1(R)"            "gL(L)"            
#>   [7] "PB(L9)"            "PB(R3)"            "a'L(L)"           
#>  [10] "NO3(R)"            "SAD(-AMMC)"        "POC"              
#>  [13] "PB(L2)"            "AL-DM2(R)"         "AL-VM2(R)"        
#>  [16] "AL-VP1m(R)"        "EPA(L)"            "AL-VM5v(R)"       
#>  [19] "AL(L)"             "NO2(R)"            "AL-VP3(R)"        
#>  [22] "GA(R)"             "SCL(R)"            "EBr3d"            
#>  [25] "GOR(R)"            "AL-DM3(R)"         "AL-DM5(R)"        
#>  [28] "PB"                "a'1(R)"            "PB(L5)"           
#>  [31] "NO1(L)"            "AL-DC4(R)"         "AL-VA1v(R)"       
#>  [34] "PRW"               "FB-column3"        "AMMC"             
#>  [37] "FBl5"              "AL-DL2v(R)"        "AL-VC5(R)"        
#>  [40] "PB(R7)"            "CRE(R)"            "AL-VM4(R)"        
#>  [43] "AL-DC4(L)"         "b1(R)"             "AOT(R)"           
#>  [46] "IB"                "AL-DL3(R)"         "AL-VM1(R)"        
#>  [49] "CA(L)"             "GOR(L)"            "mALT(R)"          
#>  [52] "EB"                "AL(R)"             "b'2(R)"           
#>  [55] "PB(R9)"            "AL-DL4(R)"         "SPS(R)"           
#>  [58] "PB(L1)"            "VMNP"              "CAN(R)"           
#>  [61] "LOP(R)"            "FBl9"              "AL-DA3(R)"        
#>  [64] "AL-DC1(R)"         "AL-VC1(R)"         "AL-VL1(R)"        
#>  [67] "AB(L)"             "vACA(R)"           "LAL(L)"           
#>  [70] "GF(R)"             "NO"                "AL-DA2(L)"        
#>  [73] "ROB(R)"            "AL-VA7m(R)"        "bL(L)"            
#>  [76] "VES(R)"            "AOTU(R)"           "AL-DL4(L)"        
#>  [79] "FBl4"              "SMP(R)"            "b2(R)"            
#>  [82] "AL-DA3(L)"         "AL-DM4(L)"         "AL-DC2(L)"        
#>  [85] "ATL(L)"            "AL-VM7d(L)"        "INP"              
#>  [88] "lACA(R)"           "AL-DM2(L)"         "SIP(R)"           
#>  [91] "AL-VC3m(R)"        "hemibrain"         "AL-DA4l(R)"       
#>  [94] "AL-DL2d(R)"        "AL-VP2(R)"         "CRE(-ROB,-RUB)(R)"
#>  [97] "FLA(R)"            "EBr3pw"            "ATL(R)"           
#> [100] "IPS(R)"            "RUB(L)"            "AL-DM1(R)"        
#> [103] "FBl1"              "AB(R)"             "AME(R)"           
#> [106] "NO2(L)"            "LAL(-GA)(R)"       "EPA(R)"           
#> [109] "AL-DM6(L)"         "NO(R)"             "g5(R)"            
#> [112] "LAL(R)"            "MB(R)"             "EBr6"             
#> [115] "AL-VA2(R)"         "a'2(R)"            "g4(R)"            
#> [118] "PB(L7)"            "AL-DM6(R)"         "SIP(L)"           
#> [121] "b'L(R)"            "FBl2"              "AL-VA5(R)"        
#> [124] "RUB(R)"            "SAD"               "PED(R)"           
#> [127] "ME(R)"             "AL-VL2p(R)"        "FBl8"             
#> [130] "PB(R2)"            "PLP(R)"            "AL-VA4(R)"        
#> [133] "AL-V(R)"           "PENP"              "a2(R)"            
#> [136] "aL(R)"             "EBr5"              "CA(R)"            
#> [139] "AL-DM1(L)"         "LO(R)"             "AL-DA4m(R)"       
#> [142] "AL-VC2(R)"         "AL-DM3(L)"         "GNG"              
#> [145] "EBr1"              "BU(L)"             "EBr3am"           
#> [148] "a1(R)"             "NO3(L)"            "PB(R5)"           
#> [151] "AL-DP1m(L)"        "VES(L)"            "AL-DP1l(R)"       
#> [154] "SMP(L)"            "SPS(L)"            "LX(R)"            
#> [157] "AL-DA4m(L)"        "PVLP(R)"           "FBl6"             
#> [160] "PB(R1)"            "CRE(L)"            "SLP(R)"           
#> [163] "AL-VA1d(R)"        "dACA(R)"           "ICL(L)"           
#> [166] "AL-DC2(R)"         "WED(R)"            "LX(L)"            
#> [169] "b'L(L)"            "AL-VM7d(R)"        "b'1(R)"           
#> [172] "AL-DC1(L)"         "AL-DL5(L)"         "AL-VM7v(L)"       
#> [175] "AL-VC4(R)"         "PB(R6)"            "SNP(R)"           
#> [178] "PB(R4)"            "AL-VA3(R)"         "MB(L)"            
#> [181] "g2(R)"             "PB(L8)"            "BU(R)"            
#> [184] "bL(R)"             "a'L(R)"            "AL-VM5d(R)"       
#> [187] "AL-DM5(L)"         "AL-VA6(L)"         "LH(R)"            
#> [190] "AL-DL1(R)"         "AL-VA6(R)"         "CX"               
#> [193] "PB(L6)"            "CRE(-RUB)(L)"      "MB(+ACA)(R)"      
#> [196] "PB(L4)"            "AL-VP4(R)"         "SCL(L)"           
#> [199] "AL-DC3(R)"         "AL-D(R)"           "GC"               
#> [202] "FBl7"              "AL-DA1(R)"         "AL-VL2a(R)"       
#> [205] "AL-VM3(R)"         "aL(L)"             "AL-VC3l(R)"       
#> [208] "AL-D(L)"           "AL-VP1d(R)"        "g1(R)"            
#> [211] "g3(R)"             "FB"                "FBl3"             
#> [214] "a3(R)"             "AL-DL5(R)"         "AL-DM4(R)"        
#> [217] "AVLP(R)"           "AL-VP1l(R)"        "PB(L3)"           
#> [220] "gL(R)"             "ICL(R)"            "VLNP(R)"          
#> [223] "a'3(R)"            "PB(R8)"            "EBr2r4"           
#> [226] "SNP(L)"            "OL(R)"             "AL-DP1m(R)"       
#> [229] "AL-VA7l(R)"        "AL-VM7v(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "NO3(L)"            "MB(+ACA)(R)"       "AL-VC3l(R)"       
#>   [4] "AL-VM7d(R)"        "AOT(R)"            "CA(L)"            
#>   [7] "AL-DM5(R)"         "vACA(R)"           "PB(L5)"           
#>  [10] "AB(R)"             "PB(R7)"            "PLP(R)"           
#>  [13] "AL-DM6(L)"         "AL-VM4(R)"         "SPS(R)"           
#>  [16] "VMNP"              "PB(L9)"            "AL-VP1d(R)"       
#>  [19] "hemibrain"         "AL-DC1(L)"         "RUB(L)"           
#>  [22] "AL-VM5v(R)"        "aL(R)"             "AL-DM3(R)"        
#>  [25] "AL-DP1l(R)"        "AL-DM1(R)"         "a'L(R)"           
#>  [28] "MB(R)"             "AL-DL1(R)"         "EBr6"             
#>  [31] "AL-DL3(R)"         "AL-DM2(R)"         "AL-VC5(R)"        
#>  [34] "CRE(-ROB,-RUB)(R)" "ME(R)"             "PB(R3)"           
#>  [37] "AL-DA1(R)"         "PVLP(R)"           "LH(R)"            
#>  [40] "AL(R)"             "AL-DM4(R)"         "AL-DC3(R)"        
#>  [43] "AL(L)"             "a3(R)"             "b'L(L)"           
#>  [46] "NO(L)"             "GA(R)"             "SNP(R)"           
#>  [49] "NO(R)"             "AL-VM3(R)"         "LAL(L)"           
#>  [52] "IB"                "a1(R)"             "PB(L2)"           
#>  [55] "mALT(R)"           "LO(R)"             "FBl4"             
#>  [58] "AL-DC2(L)"         "AL-DM5(L)"         "CAN(R)"           
#>  [61] "PB(R6)"            "AOTU(R)"           "FB"               
#>  [64] "b'1(R)"            "PB"                "gL(L)"            
#>  [67] "AVLP(R)"           "NO3(R)"            "AL-DA3(R)"        
#>  [70] "AL-DM1(L)"         "VES(R)"            "FBl7"             
#>  [73] "AL-DP1m(R)"        "AL-VC4(R)"         "VLNP(R)"          
#>  [76] "PENP"              "PRW"               "PB(L8)"           
#>  [79] "GC"                "EB"                "LOP(R)"           
#>  [82] "EPA(L)"            "GOR(R)"            "AL-VP1m(R)"       
#>  [85] "b2(R)"             "AL-DL4(L)"         "AL-DP1m(L)"       
#>  [88] "g1(R)"             "AL-DC1(R)"         "AL-VA7m(R)"       
#>  [91] "g3(R)"             "AL-VC2(R)"         "AL-VL2a(R)"       
#>  [94] "EBr3am"            "SPS(L)"            "g5(R)"            
#>  [97] "AL-VA5(R)"         "AL-VA2(R)"         "AL-VC3m(R)"       
#> [100] "mALT(L)"           "SMP(L)"            "SAD"              
#> [103] "FBl8"              "EBr3pw"            "FBl9"             
#> [106] "LX(L)"             "AL-D(R)"           "BU(L)"            
#> [109] "AL-DC4(L)"         "ROB(R)"            "AL-VA4(R)"        
#> [112] "b'L(R)"            "FBl1"              "AL-DA2(R)"        
#> [115] "ATL(R)"            "AL-VP5(R)"         "AL-DA2(L)"        
#> [118] "PB(R2)"            "AL-DL2d(R)"        "EBr1"             
#> [121] "FLA(R)"            "SIP(R)"            "g2(R)"            
#> [124] "PB(R5)"            "AL-VL1(R)"         "AL-VM7v(R)"       
#> [127] "AL-V(R)"           "dACA(R)"           "SNP(L)"           
#> [130] "a'3(R)"            "lACA(R)"           "LAL(R)"           
#> [133] "IPS(R)"            "AL-VP2(R)"         "AL-DL4(R)"        
#> [136] "a'1(R)"            "PB(L7)"            "AL-DL2v(R)"       
#> [139] "EBr2r4"            "AME(R)"            "BU(R)"            
#> [142] "bL(R)"             "VES(L)"            "LAL(-GA)(R)"      
#> [145] "CRE(R)"            "AL-DA3(L)"         "AL-DL5(R)"        
#> [148] "AL-DA4l(R)"        "AL-DC2(R)"         "gL(R)"            
#> [151] "b'2(R)"            "AL-VC1(R)"         "bL(L)"            
#> [154] "NO2(R)"            "a'L(L)"            "SMP(R)"           
#> [157] "SCL(L)"            "PED(R)"            "LX(R)"            
#> [160] "AL-D(L)"           "WED(R)"            "EBr3d"            
#> [163] "AL-VL2p(R)"        "AL-DA4m(R)"        "NO2(L)"           
#> [166] "a2(R)"             "a'2(R)"            "GNG"              
#> [169] "AL-VA3(R)"         "AB(L)"             "NO1(L)"           
#> [172] "NO"                "AL-VA1d(R)"        "INP"              
#> [175] "AL-DC4(R)"         "FB-column3"        "FBl5"             
#> [178] "ICL(R)"            "AL-VM1(R)"         "PB(R1)"           
#> [181] "SLP(R)"            "AL-VA7l(R)"        "GF(R)"            
#> [184] "FBl6"              "AL-DM6(R)"         "AL-VM5d(R)"       
#> [187] "AL-VP1l(R)"        "CX"                "SIP(L)"           
#> [190] "PB(R8)"            "PB(R9)"            "AL-DM3(L)"        
#> [193] "AL-VM7d(L)"        "AMMC"              "ICL(L)"           
#> [196] "SAD(-AMMC)"        "NO1(R)"            "AL-DL5(L)"        
#> [199] "CRE(-RUB)(L)"      "AL-DM2(L)"         "PB(L6)"           
#> [202] "AL-VA6(R)"         "POC"               "MB(L)"            
#> [205] "PB(R4)"            "AL-VA1v(R)"        "PB(L1)"           
#> [208] "AL-VP4(R)"         "RUB(R)"            "g4(R)"            
#> [211] "b1(R)"             "AL-DA4m(L)"        "AL-DM4(L)"        
#> [214] "PB(L4)"            "AL-VM2(R)"         "PB(L3)"           
#> [217] "AL-VA6(L)"         "FBl3"              "aL(L)"            
#> [220] "EBr5"              "EPA(R)"            "CA(R)"            
#> [223] "FBl2"              "OL(R)"             "AL-VP3(R)"        
#> [226] "SCL(R)"            "AL-VM7v(L)"        "CRE(L)"           
#> [229] "GOR(L)"            "ATL(L)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                               
#> [1,] "CONNECT" "/api"                             
#> [2,] "POST"    "/api/npexplorer/simpleconnections"
#> [3,] "POST"    "/api/v:ver/raw/cypher/transaction"
#> [4,] "GET"     "/api/v:ver/dbmeta/datasets"       
#> [5,] "POST"    "/api/custom/custom"               
#> [6,] "GET"     "/profile"                         
#>      name                                                                                             
#> [1,] "github.com/labstack/echo.glob..func1"                                                           
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"                  
#> [6,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                         

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }