Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "WED(R)"            "PVLP(R)"           "CX"               
#>   [4] "NO2(L)"            "EB"                "a'1(R)"           
#>   [7] "AVLP(R)"           "VLNP(R)"           "GF(R)"            
#>  [10] "a3(R)"             "SPS(R)"            "EBr6"             
#>  [13] "PLP(R)"            "GNG"               "RUB(R)"           
#>  [16] "PB(R7)"            "mALT(R)"           "EBr2r4"           
#>  [19] "IPS(R)"            "FLA(R)"            "a'2(R)"           
#>  [22] "PB(R9)"            "EBr3pw"            "GOR(R)"           
#>  [25] "IB"                "NO1(L)"            "g5(R)"            
#>  [28] "LH(R)"             "PB(R2)"            "LAL(R)"           
#>  [31] "OL(R)"             "LOP(R)"            "PB(L9)"           
#>  [34] "VES(R)"            "NO2(R)"            "LAL(L)"           
#>  [37] "a2(R)"             "SAD"               "CRE(R)"           
#>  [40] "PB"                "gL(L)"             "g3(R)"            
#>  [43] "SCL(R)"            "CA(R)"             "EPA(L)"           
#>  [46] "hemibrain"         "aL(R)"             "VMNP"             
#>  [49] "g4(R)"             "ICL(L)"            "LO(R)"            
#>  [52] "AL(R)"             "EBr5"              "NO1(R)"           
#>  [55] "PB(L6)"            "ATL(R)"            "a1(R)"            
#>  [58] "bL(L)"             "NO3(L)"            "NO3(R)"           
#>  [61] "FBl8"              "CAN(R)"            "NO(L)"            
#>  [64] "PB(R6)"            "g1(R)"             "FBl2"             
#>  [67] "VES(L)"            "SLP(R)"            "PB(L3)"           
#>  [70] "AOTU(R)"           "b2(R)"             "g2(R)"            
#>  [73] "SNP(R)"            "AME(R)"            "CA(L)"            
#>  [76] "gL(R)"             "FBl4"              "FBl7"             
#>  [79] "PB(R3)"            "LX(L)"             "RUB(L)"           
#>  [82] "NO"                "FBl1"              "vACA(R)"          
#>  [85] "FB"                "dACA(R)"           "PB(R1)"           
#>  [88] "PB(R5)"            "SCL(L)"            "MB(L)"            
#>  [91] "SPS(L)"            "PB(R8)"            "a'L(R)"           
#>  [94] "SNP(L)"            "b1(R)"             "AB(R)"            
#>  [97] "a'L(L)"            "SMP(L)"            "PB(L1)"           
#> [100] "ME(R)"             "PB(L7)"            "SAD(-AMMC)"       
#> [103] "FBl3"              "AOT(R)"            "CRE(-ROB,-RUB)(R)"
#> [106] "PB(L8)"            "AMMC"              "GA(R)"            
#> [109] "EPA(R)"            "NO(R)"             "SMP(R)"           
#> [112] "EBr3am"            "b'2(R)"            "BU(R)"            
#> [115] "bL(R)"             "b'L(L)"            "mALT(L)"          
#> [118] "SIP(L)"            "ROB(R)"            "aL(L)"            
#> [121] "POC"               "b'1(R)"            "PB(L2)"           
#> [124] "FBl6"              "PB(R4)"            "PENP"             
#> [127] "LAL(-GA)(R)"       "b'L(R)"            "BU(L)"            
#> [130] "EBr3d"             "FBl5"              "MB(R)"            
#> [133] "PB(L5)"            "PRW"               "lACA(R)"          
#> [136] "PED(R)"            "EBr1"              "CRE(L)"           
#> [139] "ATL(L)"            "INP"               "AB(L)"            
#> [142] "GC"                "LX(R)"             "SIP(R)"           
#> [145] "MB(+ACA)(R)"       "FBl9"              "ICL(R)"           
#> [148] "GOR(L)"            "AL(L)"             "PB(L4)"           
#> [151] "a'3(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "a2(R)"             "AL-DL4(L)"         "AL-DL2v(R)"       
#>   [4] "AL-DM1(R)"         "AMMC"              "AL(L)"            
#>   [7] "AL-DA2(L)"         "PVLP(R)"           "SNP(R)"           
#>  [10] "AL-VA4(R)"         "CRE(L)"            "NO"               
#>  [13] "AL-D(R)"           "AL-VC4(R)"         "NO3(R)"           
#>  [16] "g2(R)"             "AL-DC4(L)"         "g5(R)"            
#>  [19] "AL-VM4(R)"         "AL-V(R)"           "gL(R)"            
#>  [22] "AL-DM2(R)"         "AL-VP1d(R)"        "lACA(R)"          
#>  [25] "PB(R8)"            "AL-DC2(L)"         "VES(R)"           
#>  [28] "FBl6"              "gL(L)"             "AL-DM2(L)"        
#>  [31] "SAD(-AMMC)"        "PB(L4)"            "AL-VM2(R)"        
#>  [34] "GF(R)"             "AB(R)"             "AL-VC3l(R)"       
#>  [37] "AL-VL2p(R)"        "GC"                "b2(R)"            
#>  [40] "AME(R)"            "RUB(L)"            "RUB(R)"           
#>  [43] "LX(R)"             "PB(R9)"            "ROB(R)"           
#>  [46] "PB(L6)"            "SMP(L)"            "dACA(R)"          
#>  [49] "PB(L5)"            "EBr2r4"            "FBl1"             
#>  [52] "AL-DL4(R)"         "PB(R5)"            "vACA(R)"          
#>  [55] "a1(R)"             "PRW"               "EB"               
#>  [58] "GOR(R)"            "AL-VM3(R)"         "PB(L1)"           
#>  [61] "OL(R)"             "AL(R)"             "EPA(R)"           
#>  [64] "AL-VA2(R)"         "AL-VL1(R)"         "VLNP(R)"          
#>  [67] "PB(L9)"            "AL-DM6(R)"         "AL-VA1v(R)"       
#>  [70] "aL(R)"             "ICL(R)"            "AL-VA1d(R)"       
#>  [73] "AL-VP5(R)"         "FBl3"              "INP"              
#>  [76] "MB(L)"             "EBr6"              "PB(R3)"           
#>  [79] "AL-DC4(R)"         "LAL(L)"            "g1(R)"            
#>  [82] "VES(L)"            "hemibrain"         "NO3(L)"           
#>  [85] "a'L(L)"            "AL-VM5v(R)"        "AL-DA4m(L)"       
#>  [88] "ATL(R)"            "LX(L)"             "SMP(R)"           
#>  [91] "CA(L)"             "a'2(R)"            "AL-DC1(L)"        
#>  [94] "AL-DM3(L)"         "FBl9"              "AL-VM1(R)"        
#>  [97] "AL-VM7v(R)"        "PB(L2)"            "LH(R)"            
#> [100] "AL-DM3(R)"         "LO(R)"             "CA(R)"            
#> [103] "AL-DP1m(R)"        "AL-VM5d(R)"        "AL-VP1l(R)"       
#> [106] "SPS(R)"            "AL-DL2d(R)"        "AL-VP1m(R)"       
#> [109] "NO2(L)"            "AL-DC1(R)"         "ME(R)"            
#> [112] "AL-VM7v(L)"        "EBr3pw"            "FBl2"             
#> [115] "aL(L)"             "AL-DC3(R)"         "AL-VP4(R)"        
#> [118] "CRE(-RUB)(L)"      "AOTU(R)"           "PB(R1)"           
#> [121] "CAN(R)"            "AL-VA5(R)"         "SPS(L)"           
#> [124] "AL-DM1(L)"         "NO(R)"             "bL(L)"            
#> [127] "FBl5"              "NO(L)"             "AL-VA6(R)"        
#> [130] "SAD"               "PB(L8)"            "PLP(R)"           
#> [133] "AL-VM7d(R)"        "AL-DM6(L)"         "SIP(R)"           
#> [136] "MB(R)"             "AB(L)"             "CRE(-ROB,-RUB)(R)"
#> [139] "PB(R7)"            "CRE(R)"            "PB"               
#> [142] "PB(R2)"            "b1(R)"             "AL-DA4m(R)"       
#> [145] "LOP(R)"            "FLA(R)"            "PED(R)"           
#> [148] "EBr5"              "AL-DL5(L)"         "g3(R)"            
#> [151] "BU(L)"             "AL-D(L)"           "AL-VA7m(R)"       
#> [154] "CX"                "EBr1"              "b'2(R)"           
#> [157] "AL-DM4(L)"         "AL-DP1m(L)"        "SLP(R)"           
#> [160] "AL-DA3(R)"         "AL-DA4l(R)"        "AL-VA7l(R)"       
#> [163] "FBl7"              "AL-VP2(R)"         "PB(L3)"           
#> [166] "LAL(-GA)(R)"       "mALT(R)"           "bL(R)"            
#> [169] "a3(R)"             "LAL(R)"            "AL-DP1l(R)"       
#> [172] "a'3(R)"            "FB"                "ICL(L)"           
#> [175] "AVLP(R)"           "GOR(L)"            "AL-VL2a(R)"       
#> [178] "PB(R6)"            "FB-column3"        "AL-DC2(R)"        
#> [181] "AL-VC5(R)"         "b'1(R)"            "a'L(R)"           
#> [184] "NO2(R)"            "b'L(L)"            "AL-DM5(R)"        
#> [187] "SIP(L)"            "AL-VA6(L)"         "PB(L7)"           
#> [190] "AL-DM4(R)"         "IB"                "AL-DA3(L)"        
#> [193] "AL-DA1(R)"         "AL-VA3(R)"         "SNP(L)"           
#> [196] "FBl4"              "AL-VC3m(R)"        "g4(R)"            
#> [199] "AL-DL3(R)"         "AL-VP3(R)"         "AOT(R)"           
#> [202] "NO1(L)"            "IPS(R)"            "GNG"              
#> [205] "ATL(L)"            "NO1(R)"            "EBr3d"            
#> [208] "AL-VM7d(L)"        "AL-DL5(R)"         "EBr3am"           
#> [211] "SCL(R)"            "BU(R)"             "FBl8"             
#> [214] "AL-DL1(R)"         "AL-VC1(R)"         "AL-VC2(R)"        
#> [217] "SCL(L)"            "mALT(L)"           "MB(+ACA)(R)"      
#> [220] "EPA(L)"            "PB(R4)"            "WED(R)"           
#> [223] "PENP"              "AL-DA2(R)"         "b'L(R)"           
#> [226] "VMNP"              "a'1(R)"            "GA(R)"            
#> [229] "AL-DM5(L)"         "POC"              
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.2`
#> $`hemibrain:v1.2`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2`$ROIs
#>   [1] "mALT(R)"           "AL-DM3(L)"         "FBl5"             
#>   [4] "AL-VA7l(R)"        "PRW"               "CRE(R)"           
#>   [7] "EBr5"              "b1(R)"             "NO2(R)"           
#>  [10] "LAL(L)"            "SIP(L)"            "NO"               
#>  [13] "a'L(L)"            "AL-DL2v(R)"        "b'1(R)"           
#>  [16] "a'2(R)"            "AL-DP1m(L)"        "AL-DM1(R)"        
#>  [19] "AL-VC5(R)"         "SCL(L)"            "GA(R)"            
#>  [22] "OL(R)"             "AL-DM6(R)"         "AL-VP1m(R)"       
#>  [25] "dACA(R)"           "AL-VM7v(L)"        "CA(R)"            
#>  [28] "AL-VA2(R)"         "RUB(R)"            "PB"               
#>  [31] "AL-VM7d(R)"        "AL-VP5(R)"         "g5(R)"            
#>  [34] "AL-DA1(R)"         "AL-DA3(R)"         "a'3(R)"           
#>  [37] "NO1(L)"            "GNG"               "AL-DL5(R)"        
#>  [40] "VLNP(R)"           "CX"                "FB"               
#>  [43] "b'L(R)"            "EB"                "EPA(L)"           
#>  [46] "AL-DC1(R)"         "PB(L1)"            "PB(L7)"           
#>  [49] "POC"               "AL-DM6(L)"         "VES(L)"           
#>  [52] "PB(L2)"            "CRE(-ROB,-RUB)(R)" "PB(R6)"           
#>  [55] "NO1(R)"            "LH(R)"             "bL(R)"            
#>  [58] "AL-DL1(R)"         "PLP(R)"            "PB(R1)"           
#>  [61] "VES(R)"            "AL-VL2a(R)"        "PB(R8)"           
#>  [64] "AL-DA2(L)"         "PB(L9)"            "AL-DC4(R)"        
#>  [67] "SAD"               "AL-DL5(L)"         "ICL(L)"           
#>  [70] "AB(L)"             "SIP(R)"            "PB(L4)"           
#>  [73] "MB(R)"             "CA(L)"             "PB(R3)"           
#>  [76] "AL-DP1m(R)"        "aL(L)"             "AL-VP4(R)"        
#>  [79] "ATL(L)"            "BU(L)"             "ROB(R)"           
#>  [82] "AL-DL2d(R)"        "AL-VP1l(R)"        "IB"               
#>  [85] "PVLP(R)"           "SLP(R)"            "PB(L8)"           
#>  [88] "PB(R7)"            "LX(R)"             "ICL(R)"           
#>  [91] "IPS(R)"            "b'L(L)"            "AL-VC1(R)"        
#>  [94] "LOP(R)"            "GF(R)"             "PB(L3)"           
#>  [97] "SNP(R)"            "g4(R)"             "FBl7"             
#> [100] "INP"               "AL-VA5(R)"         "PB(R5)"           
#> [103] "NO(L)"             "ME(R)"             "AL-DC2(L)"        
#> [106] "EBr3pw"            "AL-DM4(R)"         "FBl2"             
#> [109] "GOR(R)"            "EBr3am"            "vACA(R)"          
#> [112] "AOTU(R)"           "AL-DC1(L)"         "AL-DL4(L)"        
#> [115] "PB(L6)"            "AL-VP2(R)"         "SMP(L)"           
#> [118] "a2(R)"             "ATL(R)"            "AL-VA1d(R)"       
#> [121] "b'2(R)"            "AL-DA4m(L)"        "hemibrain"        
#> [124] "AL(L)"             "CRE(-RUB)(L)"      "FB-column3"       
#> [127] "RUB(L)"            "PB(R2)"            "g3(R)"            
#> [130] "MB(+ACA)(R)"       "FBl9"              "LAL(R)"           
#> [133] "AL-VL2p(R)"        "FBl6"              "AL-DL4(R)"        
#> [136] "AL-V(R)"           "PB(R9)"            "AB(R)"            
#> [139] "AL-VP1d(R)"        "lACA(R)"           "bL(L)"            
#> [142] "FBl1"              "GOR(L)"            "NO3(R)"           
#> [145] "CRE(L)"            "AL-VM5d(R)"        "SPS(R)"           
#> [148] "AL-VA6(R)"         "AL-VC3m(R)"        "AL-DC4(L)"        
#> [151] "AL-VM2(R)"         "AL-VM5v(R)"        "AL-VM7v(R)"       
#> [154] "FBl3"              "LX(L)"             "a3(R)"            
#> [157] "AL-DM5(R)"         "AL-VM3(R)"         "WED(R)"           
#> [160] "SCL(R)"            "EBr1"              "AL-VM1(R)"        
#> [163] "b2(R)"             "AL-DM2(L)"         "AL-VA1v(R)"       
#> [166] "AL-DM2(R)"         "AL-VC3l(R)"        "PENP"             
#> [169] "gL(L)"             "AL-DM1(L)"         "AL-VA6(L)"        
#> [172] "AME(R)"            "AL-DM3(R)"         "AL-VA4(R)"        
#> [175] "AL-VA7m(R)"        "a'L(R)"            "FBl4"             
#> [178] "FBl8"              "AL-DC2(R)"         "AL-DC3(R)"        
#> [181] "AMMC"              "GC"                "EBr2r4"           
#> [184] "SAD(-AMMC)"        "AL-DA4m(R)"        "NO(R)"            
#> [187] "AL-DA4l(R)"        "AL-VC2(R)"         "EPA(R)"           
#> [190] "AL-DA2(R)"         "AL-DL3(R)"         "AL-DM4(L)"        
#> [193] "AVLP(R)"           "AL(R)"             "AL-VM4(R)"        
#> [196] "AL-D(L)"           "NO3(L)"            "LO(R)"            
#> [199] "aL(R)"             "AL-VM7d(L)"        "SMP(R)"           
#> [202] "VMNP"              "CAN(R)"            "AL-DP1l(R)"       
#> [205] "AOT(R)"            "PB(L5)"            "NO2(L)"           
#> [208] "EBr6"              "SPS(L)"            "EBr3d"            
#> [211] "gL(R)"             "AL-DA3(L)"         "a'1(R)"           
#> [214] "PB(R4)"            "AL-VA3(R)"         "AL-VC4(R)"        
#> [217] "PED(R)"            "g2(R)"             "SNP(L)"           
#> [220] "AL-D(R)"           "MB(L)"             "BU(R)"            
#> [223] "AL-DM5(L)"         "mALT(L)"           "a1(R)"            
#> [226] "LAL(-GA)(R)"       "g1(R)"             "FLA(R)"           
#> [229] "AL-VL1(R)"         "AL-VP3(R)"        
#> 
#> $`hemibrain:v1.2`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method  path                                     
#> [1,] "GET"   "/api"                                   
#> [2,] "GET"   "/api/v:ver/npexplorer/simpleconnections"
#> [3,] "TRACE" "/api"                                   
#> [4,] "GET"   "/api/v:ver/npexplorer/roisinneuron"     
#> [5,] "POST"  "/api/npexplorer/roisinneuron"           
#> [6,] "GET"   "/api/npexplorer/autapses"               
#>      name                                                                                        
#> [1,] "github.com/labstack/echo.glob..func1"                                                      
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [3,] "github.com/labstack/echo.glob..func1"                                                      
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"    

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }