Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "FBl2"              "CX"                "PB(R9)"           
#>   [4] "AL-VC3m(R)"        "g2(R)"             "MB(R)"            
#>   [7] "POC"               "PB(R3)"            "EBr3d"            
#>  [10] "FBl5"              "LAL(R)"            "AL-VC4(R)"        
#>  [13] "AL-VL2a(R)"        "GA(R)"             "SNP(R)"           
#>  [16] "gL(R)"             "b1(R)"             "AL-VP1m(R)"       
#>  [19] "NO3(R)"            "SIP(L)"            "INP"              
#>  [22] "AL-DM1(R)"         "SMP(L)"            "MB(L)"            
#>  [25] "ATL(R)"            "aL(L)"             "SLP(R)"           
#>  [28] "AL-V(R)"           "hemibrain"         "EBr5"             
#>  [31] "GNG"               "AL-VM5v(R)"        "AL-VP5(R)"        
#>  [34] "PB(L3)"            "AL-DP1l(R)"        "PB(R6)"           
#>  [37] "IB112"             "AL-VA1d(R)"        "PB(R5)"           
#>  [40] "vACA(R)"           "NO"                "b2(R)"            
#>  [43] "AL-VA5(R)"         "GOR(L)"            "AL-DM2(R)"        
#>  [46] "AL-VC5(R)"         "a'3(R)"            "FB-column3"       
#>  [49] "PB(L1)"            "AL-VM7v(L)"        "AB(L)"            
#>  [52] "SPS(R)"            "mALT(L)"           "AL-DM5(R)"        
#>  [55] "PB(R8)"            "AL-DM1(L)"         "AL-VM4(R)"        
#>  [58] "ATL(L)"            "lACA(R)"           "dACA(R)"          
#>  [61] "AL-DM3(L)"         "CRE(-RUB)(L)"      "ICL(L)"           
#>  [64] "VES(R)"            "FBl7"              "b'L(L)"           
#>  [67] "PED(R)"            "IB"                "PVLP(R)"          
#>  [70] "OL(R)"             "bL(L)"             "AL-DA1(R)"        
#>  [73] "AL-VA6(R)"         "FBl8"              "FBl1"             
#>  [76] "PB(R4)"            "AL-DA3(R)"         "EPA(R)"           
#>  [79] "AL-DM5(L)"         "AL-VP4(R)"         "SCL(R)"           
#>  [82] "SAD(-AMMC)"        "PB(L4)"            "AL-DA4m(R)"       
#>  [85] "EB"                "LH(R)"             "FB"               
#>  [88] "PB(L6)"            "PB(L9)"            "AL-DC2(R)"        
#>  [91] "AMMC"              "AL-DL5(L)"         "AL-DL1(R)"        
#>  [94] "NO1(L)"            "VMNP"              "AL-VM2(R)"        
#>  [97] "mALT(R)"           "PB(R7)"            "LAL(-GA)(R)"      
#> [100] "FBl4"              "LX(L)"             "AL-VM5d(R)"       
#> [103] "AOT(R)"            "RUB(R)"            "b'2(R)"           
#> [106] "AL-DP1m(L)"        "SIP(R)"            "AL-DC1(R)"        
#> [109] "AL-VM7d(R)"        "PENP"              "AL-VC2(R)"        
#> [112] "AL-VL2p(R)"        "EBr1"              "AL-DL5(R)"        
#> [115] "AL-VA2(R)"         "PB(L2)"            "b'L(R)"           
#> [118] "AL-DC4(L)"         "IPS(R)"            "VES(L)"           
#> [121] "AL-DC2(L)"         "FBl9"              "AL(L)"            
#> [124] "AL-VM1(R)"         "FLA(R)"            "PB(L8)"           
#> [127] "AL-DL4(L)"         "AL-DM3(R)"         "GF(R)"            
#> [130] "PRW"               "a'1(R)"            "AL-DM6(R)"        
#> [133] "a2(R)"             "PB"                "CRE(L)"           
#> [136] "AL-VM3(R)"         "NO1(R)"            "AL-DA4m(L)"       
#> [139] "AL-D(L)"           "PB(R2)"            "AL-DP1m(R)"       
#> [142] "a'L(R)"            "AB(R)"             "AL-DM2(L)"        
#> [145] "AL-DL3(R)"         "AL-DM4(L)"         "AL-VA4(R)"        
#> [148] "NO2(R)"            "AL-DA2(R)"         "AL-DL2d(R)"       
#> [151] "CA(L)"             "SPS(L)"            "a'L(L)"           
#> [154] "NO3(L)"            "PB(L5)"            "a1(R)"            
#> [157] "AVLP(R)"           "gL(L)"             "AL-VA3(R)"        
#> [160] "b'1(R)"            "g3(R)"             "BU(R)"            
#> [163] "NO(R)"             "SNP(L)"            "AL-DL2v(R)"       
#> [166] "AL-VL1(R)"         "g1(R)"             "bL(R)"            
#> [169] "AL-VC1(R)"         "CRE(R)"            "a3(R)"            
#> [172] "AL-VA7m(R)"        "AL-VP1d(R)"        "AL-DL4(R)"        
#> [175] "BU(L)"             "ME(R)"             "AL-DC1(L)"        
#> [178] "RUB(L)"            "SCL(L)"            "GC"               
#> [181] "a'2(R)"            "FBl3"              "FBl6"             
#> [184] "AL-DC3(R)"         "AL-VP3(R)"         "EBr6"             
#> [187] "EBr3pw"            "PB(L7)"            "AME(R)"           
#> [190] "AOTU(R)"           "PB(R1)"            "CAN(R)"           
#> [193] "AL-VA1v(R)"        "LOP(R)"            "PLP(R)"           
#> [196] "CA(R)"             "g4(R)"             "ROB(R)"           
#> [199] "MB(+ACA)(R)"       "AL-VP2(R)"         "LAL(L)"           
#> [202] "ICL(R)"            "AL-DC4(R)"         "AL-VA7l(R)"       
#> [205] "EBr2r4"            "AL-D(R)"           "AL-VM7v(R)"       
#> [208] "EBr3am"            "aL(R)"             "CRE(-ROB,-RUB)(R)"
#> [211] "g5(R)"             "AL-DA4l(R)"        "SMP(R)"           
#> [214] "AL-VA6(L)"         "AL-DM6(L)"         "AL(R)"            
#> [217] "WED(R)"            "AL-VM7d(L)"        "AL-DA2(L)"        
#> [220] "LO(R)"             "GOR(R)"            "AL-VC3l(R)"       
#> [223] "LX(R)"             "AL-VP1l(R)"        "NO2(L)"           
#> [226] "SAD"               "NO(L)"             "AL-DM4(R)"        
#> [229] "EPA(L)"            "VLNP(R)"           "AL-DA3(L)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "a'L(R)"            "NO2(R)"            "aL(L)"            
#>   [4] "VLNP(R)"           "b'1(R)"            "g3(R)"            
#>   [7] "PB(R7)"            "GC"                "g2(R)"            
#>  [10] "EBr6"              "PB(L8)"            "SNP(L)"           
#>  [13] "FBl9"              "SMP(L)"            "a'1(R)"           
#>  [16] "GNG"               "SPS(R)"            "EBr3d"            
#>  [19] "bL(L)"             "FBl8"              "LX(L)"            
#>  [22] "EBr1"              "PB(R1)"            "MB(R)"            
#>  [25] "a'3(R)"            "AOTU(R)"           "CRE(L)"           
#>  [28] "PB(L7)"            "LO(R)"             "PB(L2)"           
#>  [31] "BU(R)"             "VES(L)"            "CA(L)"            
#>  [34] "EBr2r4"            "NO1(R)"            "NO3(L)"           
#>  [37] "VES(R)"            "FBl2"              "FBl4"             
#>  [40] "LAL(L)"            "gL(L)"             "SIP(R)"           
#>  [43] "FBl7"              "PB(L4)"            "AL(R)"            
#>  [46] "b'L(R)"            "SPS(L)"            "NO3(R)"           
#>  [49] "PB(L1)"            "AL(L)"             "PB(R9)"           
#>  [52] "NO(R)"             "GOR(R)"            "LOP(R)"           
#>  [55] "WED(R)"            "PB(R8)"            "INP"              
#>  [58] "SCL(L)"            "LAL(-GA)(R)"       "g5(R)"            
#>  [61] "aL(R)"             "SNP(R)"            "ROB(R)"           
#>  [64] "gL(R)"             "hemibrain"         "FBl6"             
#>  [67] "CA(R)"             "NO(L)"             "ICL(L)"           
#>  [70] "PLP(R)"            "a'2(R)"            "AVLP(R)"          
#>  [73] "PED(R)"            "EPA(R)"            "NO"               
#>  [76] "RUB(L)"            "BU(L)"             "CRE(R)"           
#>  [79] "b2(R)"             "AB(R)"             "a3(R)"            
#>  [82] "vACA(R)"           "mALT(L)"           "ATL(L)"           
#>  [85] "NO1(L)"            "MB(+ACA)(R)"       "PB(R5)"           
#>  [88] "ME(R)"             "EBr3pw"            "PB(R2)"           
#>  [91] "PRW"               "a'L(L)"            "SLP(R)"           
#>  [94] "AB(L)"             "IB"                "b'2(R)"           
#>  [97] "FLA(R)"            "CAN(R)"            "LAL(R)"           
#> [100] "SMP(R)"            "PB(L5)"            "a1(R)"            
#> [103] "dACA(R)"           "g1(R)"             "FB"               
#> [106] "PB(L9)"            "GA(R)"             "b'L(L)"           
#> [109] "POC"               "CRE(-ROB,-RUB)(R)" "SCL(R)"           
#> [112] "GF(R)"             "PB"                "EB"               
#> [115] "FBl5"              "ATL(R)"            "PB(R3)"           
#> [118] "PB(R6)"            "MB(L)"             "VMNP"             
#> [121] "LX(R)"             "FBl1"              "EPA(L)"           
#> [124] "SAD"               "NO2(L)"            "PB(R4)"           
#> [127] "lACA(R)"           "GOR(L)"            "g4(R)"            
#> [130] "PB(L6)"            "LH(R)"             "AME(R)"           
#> [133] "AMMC"              "SIP(L)"            "SAD(-AMMC)"       
#> [136] "AOT(R)"            "EBr3am"            "EBr5"             
#> [139] "a2(R)"             "mALT(R)"           "PVLP(R)"          
#> [142] "FBl3"              "bL(R)"             "PENP"             
#> [145] "ICL(R)"            "OL(R)"             "RUB(R)"           
#> [148] "CX"                "PB(L3)"            "IPS(R)"           
#> [151] "b1(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-VA6(R)"         "AL-VC1(R)"         "CRE(L)"           
#>   [4] "FBl1"              "AL-DM3(R)"         "AL-DM5(L)"        
#>   [7] "SNP(L)"            "FBl7"              "NO(L)"            
#>  [10] "AL-VA4(R)"         "ICL(L)"            "AL(R)"            
#>  [13] "PB(L3)"            "NO(R)"             "AL-DP1m(R)"       
#>  [16] "AL-DM6(L)"         "AL-VM7v(L)"        "MB(+ACA)(R)"      
#>  [19] "EPA(L)"            "AL-DL3(R)"         "AL-DM4(L)"        
#>  [22] "GNG"               "OL(R)"             "AL-VL1(R)"        
#>  [25] "AL-VM4(R)"         "CRE(-ROB,-RUB)(R)" "GOR(L)"           
#>  [28] "b'L(R)"            "PB(L7)"            "NO2(R)"           
#>  [31] "AL-DM1(R)"         "CRE(R)"            "BU(R)"            
#>  [34] "b1(R)"             "AL-DM4(R)"         "WED(R)"           
#>  [37] "AL-VM7d(L)"        "AL-DM5(R)"         "mALT(L)"          
#>  [40] "GA(R)"             "AL-DC4(R)"         "SLP(R)"           
#>  [43] "AL-DA2(R)"         "EBr6"              "AL-VM3(R)"        
#>  [46] "AL-VM7d(R)"        "b'1(R)"            "LOP(R)"           
#>  [49] "SIP(L)"            "b2(R)"             "AL-DC2(R)"        
#>  [52] "AL-DL2d(R)"        "AL-DL5(R)"         "PB(L5)"           
#>  [55] "FLA(R)"            "AL-DA1(R)"         "AL-VA5(R)"        
#>  [58] "AL-VA7m(R)"        "g4(R)"             "CRE(-RUB)(L)"     
#>  [61] "a'1(R)"            "VES(R)"            "hemibrain"        
#>  [64] "FB"                "EBr5"              "VLNP(R)"          
#>  [67] "a2(R)"             "AL-VA7l(R)"        "gL(L)"            
#>  [70] "LH(R)"             "AL-VC3m(R)"        "AL-VC4(R)"        
#>  [73] "AMMC"              "b'L(L)"            "AL-VP1d(R)"       
#>  [76] "AL-VL2a(R)"        "AL-VP1m(R)"        "LO(R)"            
#>  [79] "LX(L)"             "CAN(R)"            "AL-VA3(R)"        
#>  [82] "AL-VM2(R)"         "PRW"               "PB(L8)"           
#>  [85] "a'L(R)"            "gL(R)"             "bL(L)"            
#>  [88] "FBl5"              "EBr3d"             "AL-DL4(L)"        
#>  [91] "ICL(R)"            "AL-DC2(L)"         "NO"               
#>  [94] "g5(R)"             "dACA(R)"           "FBl2"             
#>  [97] "SNP(R)"            "FBl9"              "AL-D(R)"          
#> [100] "POC"               "g1(R)"             "NO1(R)"           
#> [103] "a'L(L)"            "AL-VM5d(R)"        "ATL(L)"           
#> [106] "PB(L1)"            "ME(R)"             "AL(L)"            
#> [109] "AL-DA4m(L)"        "AL-DP1m(L)"        "PB(R1)"           
#> [112] "AME(R)"            "aL(R)"             "PB(R9)"           
#> [115] "PB(R3)"            "AL-VP2(R)"         "NO3(R)"           
#> [118] "bL(R)"             "a3(R)"             "PB(R6)"           
#> [121] "SPS(R)"            "SMP(R)"            "PED(R)"           
#> [124] "a'2(R)"            "AL-DL2v(R)"        "AL-VP3(R)"        
#> [127] "AL-DM2(L)"         "SAD(-AMMC)"        "CA(R)"            
#> [130] "PB(R4)"            "IB"                "lACA(R)"          
#> [133] "AL-DC1(L)"         "AL-VA1v(R)"        "AL-VL2p(R)"       
#> [136] "EBr1"              "AL-DC4(L)"         "LAL(-GA)(R)"      
#> [139] "AL-D(L)"           "PVLP(R)"           "FB-column3"       
#> [142] "AL-DL1(R)"         "AL-DA2(L)"         "PB(L4)"           
#> [145] "MB(R)"             "AL-VC2(R)"         "AL-VC5(R)"        
#> [148] "CA(L)"             "AL-VA2(R)"         "AL-V(R)"          
#> [151] "AB(L)"             "RUB(R)"            "AL-DL5(L)"        
#> [154] "AL-DA4l(R)"        "SCL(R)"            "FBl6"             
#> [157] "LAL(R)"            "AL-VM5v(R)"        "AL-VP5(R)"        
#> [160] "SAD"               "PLP(R)"            "AL-VA6(L)"        
#> [163] "AOTU(R)"           "g2(R)"             "EB"               
#> [166] "PB(L9)"            "a1(R)"             "EBr3am"           
#> [169] "a'3(R)"            "SIP(R)"            "AL-DA3(R)"        
#> [172] "GC"                "PB(R2)"            "AVLP(R)"          
#> [175] "RUB(L)"            "SCL(L)"            "AL-DA3(L)"        
#> [178] "INP"               "AL-VM1(R)"         "AL-VM7v(R)"       
#> [181] "PB(R5)"            "FBl3"              "AL-VC3l(R)"       
#> [184] "CX"                "PB"                "PB(L6)"           
#> [187] "ATL(R)"            "AL-DM2(R)"         "AL-DM6(R)"        
#> [190] "PB(L2)"            "FBl8"              "AL-DC3(R)"        
#> [193] "NO2(L)"            "BU(L)"             "NO3(L)"           
#> [196] "IPS(R)"            "AL-DC1(R)"         "AL-VP1l(R)"       
#> [199] "LAL(L)"            "GF(R)"             "g3(R)"            
#> [202] "PB(R7)"            "AL-DM3(L)"         "AL-VA1d(R)"       
#> [205] "SMP(L)"            "LX(R)"             "AL-DM1(L)"        
#> [208] "AL-DP1l(R)"        "AOT(R)"            "b'2(R)"           
#> [211] "FBl4"              "aL(L)"             "EBr3pw"           
#> [214] "VES(L)"            "NO1(L)"            "VMNP"             
#> [217] "EBr2r4"            "AL-DA4m(R)"        "AL-DL4(R)"        
#> [220] "PENP"              "mALT(R)"           "PB(R8)"           
#> [223] "SPS(L)"            "AB(R)"             "vACA(R)"          
#> [226] "GOR(R)"            "MB(L)"             "EPA(R)"           
#> [229] "ROB(R)"            "AL-VP4(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "PB(L3)"            "FLA(R)"            "AL-VA6(L)"        
#>   [4] "PB(R5)"            "b'2(R)"            "AL-DL5(L)"        
#>   [7] "a'L(L)"            "PB(L2)"            "b'L(R)"           
#>  [10] "SNP(R)"            "SIP(R)"            "PB(R3)"           
#>  [13] "AL-DP1l(R)"        "AL-VP1d(R)"        "aL(R)"            
#>  [16] "NO(L)"             "AL-VC2(R)"         "EBr3am"           
#>  [19] "PB(L1)"            "FBl4"              "EPA(L)"           
#>  [22] "RUB(R)"            "ATL(R)"            "AL-VM2(R)"        
#>  [25] "CX"                "AL-DL1(R)"         "AOT(R)"           
#>  [28] "a2(R)"             "g4(R)"             "AL-DM1(L)"        
#>  [31] "AL-DP1m(L)"        "gL(R)"             "SLP(R)"           
#>  [34] "AL-DA3(L)"         "PB(L6)"            "aL(L)"            
#>  [37] "AL-DA4l(R)"        "POC"               "a1(R)"            
#>  [40] "MB(+ACA)(R)"       "CA(R)"             "AL-DM3(R)"        
#>  [43] "AL-VP1m(R)"        "SAD(-AMMC)"        "FBl3"             
#>  [46] "AB(L)"             "IB"                "AL-VC4(R)"        
#>  [49] "SPS(R)"            "SPS(L)"            "PLP(R)"           
#>  [52] "AL-VA5(R)"         "AL-VP4(R)"         "NO3(R)"           
#>  [55] "AL-DM3(L)"         "FBl7"              "AL-DA1(R)"        
#>  [58] "AL-DA3(R)"         "AL-DA4m(R)"        "NO(R)"            
#>  [61] "g3(R)"             "NO1(L)"            "FB-column3"       
#>  [64] "PB(R4)"            "AL-DL5(R)"         "AL-VA1v(R)"       
#>  [67] "ICL(R)"            "OL(R)"             "mALT(R)"          
#>  [70] "GC"                "LO(R)"             "WED(R)"           
#>  [73] "EPA(R)"            "AL-DA2(L)"         "AL-DC3(R)"        
#>  [76] "PB(R7)"            "MB(L)"             "AL-DC4(L)"        
#>  [79] "INP"               "GOR(R)"            "NO2(L)"           
#>  [82] "LAL(L)"            "FBl8"              "AME(R)"           
#>  [85] "AL-VM7v(R)"        "CA(L)"             "SIP(L)"           
#>  [88] "PB(L8)"            "AL-DA4m(L)"        "AL-D(L)"          
#>  [91] "PB(L4)"            "AL-DL3(R)"         "dACA(R)"          
#>  [94] "AL-DC2(L)"         "FBl2"              "VLNP(R)"          
#>  [97] "LOP(R)"            "AL-VA4(R)"         "PB(R8)"           
#> [100] "PB(R1)"            "AOTU(R)"           "SAD"              
#> [103] "AL-VM3(R)"         "AL-V(R)"           "AL-VP1l(R)"       
#> [106] "PENP"              "PVLP(R)"           "AL-DP1m(R)"       
#> [109] "LX(R)"             "AL-VM7d(L)"        "bL(L)"            
#> [112] "PB(L5)"            "a'2(R)"            "gL(L)"            
#> [115] "AL-VM7v(L)"        "VES(R)"            "LH(R)"            
#> [118] "AL-VM5d(R)"        "AL-DL4(R)"         "b1(R)"            
#> [121] "a3(R)"             "EBr6"              "EBr3d"            
#> [124] "GF(R)"             "hemibrain"         "AL-VC5(R)"        
#> [127] "AL-DM6(L)"         "NO3(L)"            "PB"               
#> [130] "FBl5"              "LAL(R)"            "AL-DA2(R)"        
#> [133] "AL-VM4(R)"         "PB(L9)"            "AL(R)"            
#> [136] "AL-DM4(R)"         "AL-DC1(L)"         "a'1(R)"           
#> [139] "AL-DL2d(R)"        "AL-VA6(R)"         "AL-VC3l(R)"       
#> [142] "SMP(R)"            "g1(R)"             "BU(L)"            
#> [145] "NO1(R)"            "SNP(L)"            "vACA(R)"          
#> [148] "FBl9"              "b'L(L)"            "AL-VA7m(R)"       
#> [151] "PB(R6)"            "CRE(-RUB)(L)"      "AL-VC1(R)"        
#> [154] "LAL(-GA)(R)"       "GOR(L)"            "b2(R)"            
#> [157] "EB"                "PB(R9)"            "AL-VP5(R)"        
#> [160] "PB(R2)"            "AL-DM1(R)"         "AL-VA3(R)"        
#> [163] "EBr1"              "EBr2r4"            "AL-DM6(R)"        
#> [166] "CRE(-ROB,-RUB)(R)" "VMNP"              "RUB(L)"           
#> [169] "AL-D(R)"           "ATL(L)"            "AL-DM5(R)"        
#> [172] "bL(R)"             "AL-VA2(R)"         "AL-VP3(R)"        
#> [175] "AL-DM2(L)"         "LX(L)"             "AL-VC3m(R)"       
#> [178] "BU(R)"             "ROB(R)"            "b'1(R)"           
#> [181] "mALT(L)"           "SCL(R)"            "ICL(L)"           
#> [184] "AL-VM7d(R)"        "SCL(L)"            "g2(R)"            
#> [187] "ME(R)"             "g5(R)"             "SMP(L)"           
#> [190] "CRE(R)"            "IPS(R)"            "GNG"              
#> [193] "AL-VL2a(R)"        "AL-VM1(R)"         "AL-VA7l(R)"       
#> [196] "GA(R)"             "PB(L7)"            "AB(R)"            
#> [199] "CAN(R)"            "PRW"               "NO"               
#> [202] "AVLP(R)"           "AL-VA1d(R)"        "AL-VL2p(R)"       
#> [205] "AL-DL4(L)"         "VES(L)"            "AL-DC4(R)"        
#> [208] "AL-DM2(R)"         "CRE(L)"            "EBr3pw"           
#> [211] "AL-VL1(R)"         "AL-VP2(R)"         "a'3(R)"           
#> [214] "FB"                "AL-DM5(L)"         "FBl6"             
#> [217] "MB(R)"             "AL-DC1(R)"         "AMMC"             
#> [220] "lACA(R)"           "PED(R)"            "a'L(R)"           
#> [223] "FBl1"              "NO2(R)"            "AL-DC2(R)"        
#> [226] "EBr5"              "AL(L)"             "AL-DM4(L)"        
#> [229] "AL-DL2v(R)"        "AL-VM5v(R)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                   
#> [1,] "GET"  "/api"                                 
#> [2,] "GET"  "/api/*"                               
#> [3,] "GET"  "/api/raw/keyvalue/key/:instance/:key" 
#> [4,] "GET"  "/token"                               
#> [5,] "POST" "/api/npexplorer/neuronmetavals"       
#> [6,] "POST" "/api/v:ver/npexplorer/roiconnectivity"
#>      name                                                                                          
#> [1,] "github.com/labstack/echo/v4.glob..func1"                                                     
#> [2,] "github.com/labstack/echo/v4.glob..func1"                                                     
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/raw/keyvalue.masterAPI.getKV-fm"             
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                      
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMetaVals-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }