Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "mALT(L)"           "NO3(R)"            "CA(R)"            
#>   [4] "a3(R)"             "AL-DM5(R)"         "AL-VC2(R)"        
#>   [7] "AL-VL1(R)"         "NO"                "PB(R9)"           
#>  [10] "NO3(L)"            "RUB(L)"            "NO(L)"            
#>  [13] "PVLP(R)"           "FBl7"              "NO2(R)"           
#>  [16] "AL-DC3(R)"         "AL-DC4(R)"         "AL-VP1d(R)"       
#>  [19] "a'1(R)"            "NO(R)"             "gL(R)"            
#>  [22] "FBl9"              "LX(L)"             "AL-DA4l(R)"       
#>  [25] "GOR(L)"            "MB(L)"             "FLA(R)"           
#>  [28] "SIP(R)"            "CX"                "AL-VA1d(R)"       
#>  [31] "hemibrain"         "AL-VP2(R)"         "AL-DM2(L)"        
#>  [34] "GOR(R)"            "AL-DA2(R)"         "SNP(R)"           
#>  [37] "AL-DC2(R)"         "AL-DM3(R)"         "AL-VL2p(R)"       
#>  [40] "CA(L)"             "GC"                "g2(R)"            
#>  [43] "FBl3"              "AL-VA4(R)"         "AL-VM4(R)"        
#>  [46] "PB(L5)"            "a2(R)"             "CRE(R)"           
#>  [49] "dACA(R)"           "FBl2"              "WED(R)"           
#>  [52] "LH(R)"             "AB(R)"             "AL-VM7v(R)"       
#>  [55] "LAL(L)"            "AL-VM1(R)"         "EBr5"             
#>  [58] "EB"                "AL-VM7v(L)"        "PB(R1)"           
#>  [61] "AME(R)"            "SMP(L)"            "g3(R)"            
#>  [64] "GA(R)"             "AL-DM6(L)"         "AL-VC3m(R)"       
#>  [67] "CRE(-ROB,-RUB)(R)" "NO1(R)"            "AL(L)"            
#>  [70] "AL-DM1(L)"         "IPS(R)"            "AL-DA4m(R)"       
#>  [73] "AL-VP3(R)"         "g4(R)"             "EPA(L)"           
#>  [76] "IB"                "EBr6"              "AB(L)"            
#>  [79] "ME(R)"             "AL-DP1m(L)"        "PB(L9)"           
#>  [82] "g5(R)"             "AVLP(R)"           "AL-VA5(R)"        
#>  [85] "AL-VM2(R)"         "AL-VM5d(R)"        "b'1(R)"           
#>  [88] "AOTU(R)"           "b'L(R)"            "GNG"              
#>  [91] "MB(R)"             "AL-DL4(R)"         "AL-VM7d(R)"       
#>  [94] "AL-DA4m(L)"        "AL-DA1(R)"         "AL-DM1(R)"        
#>  [97] "AL-DP1l(R)"        "ATL(L)"            "LAL(-GA)(R)"      
#> [100] "NO1(L)"            "b2(R)"             "AL-VA3(R)"        
#> [103] "PB(L1)"            "SPS(L)"            "AL-DL1(R)"        
#> [106] "VLNP(R)"           "a1(R)"             "g1(R)"            
#> [109] "FB"                "PB(R8)"            "AL-DL4(L)"        
#> [112] "OL(R)"             "AL-DL2v(R)"        "BU(R)"            
#> [115] "PB(L6)"            "FBl6"              "AL-VA1v(R)"       
#> [118] "SCL(R)"            "EBr2r4"            "AL-DA2(L)"        
#> [121] "AL-DC1(R)"         "POC"               "mALT(R)"          
#> [124] "AL-VM7d(L)"        "LAL(R)"            "AL-DM4(R)"        
#> [127] "EPA(R)"            "AL-DA3(L)"         "AL-VA6(L)"        
#> [130] "FBl4"              "PB(R4)"            "AL-DM5(L)"        
#> [133] "MB(+ACA)(R)"       "AL-V(R)"           "CAN(R)"           
#> [136] "AL-VC4(R)"         "AL-VC5(R)"         "AL-VM3(R)"        
#> [139] "AOT(R)"            "PB"                "AL-DL3(R)"        
#> [142] "AL-DC1(L)"         "PB(L3)"            "FBl1"             
#> [145] "SCL(L)"            "aL(R)"             "EBr3am"           
#> [148] "b'2(R)"            "SIP(L)"            "AL-VA7l(R)"       
#> [151] "AL-DM4(L)"         "PB(R3)"            "AL-DM6(R)"        
#> [154] "INP"               "AL-D(L)"           "AL-DL5(R)"        
#> [157] "AL-VA7m(R)"        "AL-VL2a(R)"        "AL-VM5v(R)"       
#> [160] "PENP"              "EBr1"              "LO(R)"            
#> [163] "FBl8"              "AL-D(R)"           "AL-VA2(R)"        
#> [166] "AL(R)"             "AL-VP5(R)"         "LX(R)"            
#> [169] "a'L(L)"            "SLP(R)"            "AMMC"             
#> [172] "AL-DL2d(R)"        "PLP(R)"            "AL-DL5(L)"        
#> [175] "AL-DM3(L)"         "AL-VP1l(R)"        "PB(L2)"           
#> [178] "EBr3pw"            "PB(L7)"            "CRE(L)"           
#> [181] "VES(R)"            "PB(L4)"            "LOP(R)"           
#> [184] "SNP(L)"            "b'L(L)"            "AL-VC3l(R)"       
#> [187] "BU(L)"             "AL-DC4(L)"         "PB(R7)"           
#> [190] "ROB(R)"            "FB-column3"        "PB(R2)"           
#> [193] "b1(R)"             "GF(R)"             "lACA(R)"          
#> [196] "gL(L)"             "AL-DM2(R)"         "PB(R6)"           
#> [199] "a'L(R)"            "AL-VA6(R)"         "a'3(R)"           
#> [202] "ICL(R)"            "AL-VC1(R)"         "CRE(-RUB)(L)"     
#> [205] "bL(R)"             "AL-VP1m(R)"        "AL-DC2(L)"        
#> [208] "ICL(L)"            "SAD(-AMMC)"        "ATL(R)"           
#> [211] "aL(L)"             "vACA(R)"           "SAD"              
#> [214] "PRW"               "a'2(R)"            "RUB(R)"           
#> [217] "bL(L)"             "VES(L)"            "AL-DP1m(R)"       
#> [220] "NO2(L)"            "SPS(R)"            "VMNP"             
#> [223] "IB112"             "PB(R5)"            "PED(R)"           
#> [226] "AL-DA3(R)"         "AL-VP4(R)"         "SMP(R)"           
#> [229] "PB(L8)"            "EBr3d"             "FBl5"             
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB"                "INP"               "b'L(L)"           
#>   [4] "SCL(L)"            "b'2(R)"            "GA(R)"            
#>   [7] "NO3(R)"            "VES(R)"            "CRE(-ROB,-RUB)(R)"
#>  [10] "PB(L3)"            "AOTU(R)"           "bL(R)"            
#>  [13] "GOR(R)"            "a'3(R)"            "hemibrain"        
#>  [16] "CRE(R)"            "EBr3pw"            "PB(R3)"           
#>  [19] "a1(R)"             "SPS(R)"            "NO"               
#>  [22] "FBl3"              "PB(R4)"            "NO(L)"            
#>  [25] "PB(L1)"            "AB(L)"             "PB(L2)"           
#>  [28] "CX"                "b2(R)"             "FLA(R)"           
#>  [31] "PB(L7)"            "g5(R)"             "CAN(R)"           
#>  [34] "PB(R5)"            "EB"                "g4(R)"            
#>  [37] "PB(L9)"            "a'1(R)"            "FBl5"             
#>  [40] "PB(L6)"            "PB(R6)"            "SMP(L)"           
#>  [43] "SAD"               "LX(R)"             "AL(L)"            
#>  [46] "PB(L8)"            "NO(R)"             "VLNP(R)"          
#>  [49] "NO3(L)"            "PB(R2)"            "SIP(L)"           
#>  [52] "EBr6"              "VMNP"              "SNP(R)"           
#>  [55] "bL(L)"             "vACA(R)"           "PB(R1)"           
#>  [58] "EPA(L)"            "OL(R)"             "VES(L)"           
#>  [61] "MB(R)"             "FBl2"              "LH(R)"            
#>  [64] "ATL(R)"            "LAL(L)"            "a2(R)"            
#>  [67] "g1(R)"             "ME(R)"             "FBl1"             
#>  [70] "POC"               "IB"                "NO2(L)"           
#>  [73] "g2(R)"             "FBl4"              "PB(L4)"           
#>  [76] "SMP(R)"            "AMMC"              "PB(R7)"           
#>  [79] "NO1(L)"            "PB(R8)"            "ROB(R)"           
#>  [82] "mALT(L)"           "GOR(L)"            "a'2(R)"           
#>  [85] "gL(R)"             "FBl8"              "EBr2r4"           
#>  [88] "gL(L)"             "AOT(R)"            "PENP"             
#>  [91] "EPA(R)"            "FBl6"              "RUB(R)"           
#>  [94] "CA(L)"             "lACA(R)"           "mALT(R)"          
#>  [97] "SPS(L)"            "MB(+ACA)(R)"       "aL(L)"            
#> [100] "LAL(-GA)(R)"       "BU(L)"             "FBl9"             
#> [103] "dACA(R)"           "SNP(L)"            "SAD(-AMMC)"       
#> [106] "AL(R)"             "PLP(R)"            "a'L(R)"           
#> [109] "PB(R9)"            "CRE(L)"            "LX(L)"            
#> [112] "b1(R)"             "AB(R)"             "GNG"              
#> [115] "FB"                "EBr5"              "BU(R)"            
#> [118] "ICL(L)"            "LO(R)"             "PB(L5)"           
#> [121] "PED(R)"            "aL(R)"             "IPS(R)"           
#> [124] "NO2(R)"            "LAL(R)"            "ICL(R)"           
#> [127] "LOP(R)"            "NO1(R)"            "EBr3d"            
#> [130] "FBl7"              "a'L(L)"            "PRW"              
#> [133] "b'L(R)"            "ATL(L)"            "GF(R)"            
#> [136] "PVLP(R)"           "AME(R)"            "b'1(R)"           
#> [139] "EBr3am"            "GC"                "WED(R)"           
#> [142] "a3(R)"             "SLP(R)"            "MB(L)"            
#> [145] "EBr1"              "CA(R)"             "g3(R)"            
#> [148] "SCL(R)"            "SIP(R)"            "AVLP(R)"          
#> [151] "RUB(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "GC"                "EB"                "PB(L4)"           
#>   [4] "AL-VA3(R)"         "PB(L8)"            "EPA(R)"           
#>   [7] "AL-DL5(R)"         "GA(R)"             "SAD"              
#>  [10] "NO3(R)"            "LO(R)"             "AL-DM1(R)"        
#>  [13] "PLP(R)"            "FBl9"              "AL-DA4m(R)"       
#>  [16] "RUB(R)"            "INP"               "AL-DL3(R)"        
#>  [19] "AL-VM7v(R)"        "AL-V(R)"           "VLNP(R)"          
#>  [22] "BU(L)"             "EBr2r4"            "PB(R3)"           
#>  [25] "AL-DC4(R)"         "CA(R)"             "IPS(R)"           
#>  [28] "AL-DP1l(R)"        "AL-VM3(R)"         "AL-VM4(R)"        
#>  [31] "CX"                "AL-DA2(L)"         "CRE(-RUB)(L)"     
#>  [34] "AL-VA7m(R)"        "AL-VM5v(R)"        "AL-VP3(R)"        
#>  [37] "CA(L)"             "ME(R)"             "AL-D(L)"          
#>  [40] "AL-DM6(L)"         "gL(R)"             "VES(R)"           
#>  [43] "AL-VA6(R)"         "AL-VM2(R)"         "AL-DC2(L)"        
#>  [46] "FBl1"              "LOP(R)"            "a'1(R)"           
#>  [49] "mALT(R)"           "g4(R)"             "SNP(L)"           
#>  [52] "FBl3"              "AL-DM6(R)"         "AL-VP4(R)"        
#>  [55] "g3(R)"             "NO1(R)"            "AL-DA3(L)"        
#>  [58] "GOR(R)"            "AL-DA2(R)"         "AL-VA4(R)"        
#>  [61] "FLA(R)"            "AL-DM3(R)"         "SCL(L)"           
#>  [64] "a2(R)"             "SPS(L)"            "BU(R)"            
#>  [67] "GF(R)"             "b2(R)"             "AL-DC4(L)"        
#>  [70] "SAD(-AMMC)"        "a'L(R)"            "gL(L)"            
#>  [73] "PVLP(R)"           "a3(R)"             "AL-VA2(R)"        
#>  [76] "AL-VP5(R)"         "AL-DP1m(L)"        "AVLP(R)"          
#>  [79] "RUB(L)"            "AL-DA4l(R)"        "PB(R5)"           
#>  [82] "SMP(R)"            "g5(R)"             "PB(L2)"           
#>  [85] "PRW"               "FB"                "AL-DA4m(L)"       
#>  [88] "AL-DM5(L)"         "WED(R)"            "FBl8"             
#>  [91] "AB(L)"             "MB(+ACA)(R)"       "AL-VM7d(R)"       
#>  [94] "PB(R4)"            "b'L(L)"            "aL(R)"            
#>  [97] "CRE(R)"            "b1(R)"             "SMP(L)"           
#> [100] "PB(R8)"            "AL-D(R)"           "GOR(L)"           
#> [103] "FBl2"              "aL(L)"             "AL-DA1(R)"        
#> [106] "AL-VC1(R)"         "IB"                "EBr5"             
#> [109] "AL-DM1(L)"         "ICL(L)"            "dACA(R)"          
#> [112] "AL-VC2(R)"         "LAL(L)"            "SPS(R)"           
#> [115] "AMMC"              "MB(L)"             "AL-DM3(L)"        
#> [118] "PB(R1)"            "FBl4"              "AL-VC3l(R)"       
#> [121] "PB(L3)"            "AOTU(R)"           "CRE(L)"           
#> [124] "ATL(R)"            "a'L(L)"            "AL-VP1l(R)"       
#> [127] "PB(R6)"            "AL-DM4(R)"         "SNP(R)"           
#> [130] "AL-VM7v(L)"        "AL-DL4(R)"         "AL-DM5(R)"        
#> [133] "a'3(R)"            "FBl5"              "AB(R)"            
#> [136] "AL-VP2(R)"         "AL(L)"             "PB(R2)"           
#> [139] "NO(L)"             "AL-VL2p(R)"        "mALT(L)"          
#> [142] "PED(R)"            "AL-VM7d(L)"        "bL(L)"            
#> [145] "AL-DC1(R)"         "AL-DL2v(R)"        "AL-VA7l(R)"       
#> [148] "PB(R7)"            "AL-DL5(L)"         "ICL(R)"           
#> [151] "AL-DC2(R)"         "POC"               "AOT(R)"           
#> [154] "PB(L5)"            "lACA(R)"           "FBl6"             
#> [157] "AL-DL1(R)"         "NO2(L)"            "SIP(R)"           
#> [160] "AL-VA5(R)"         "LAL(-GA)(R)"       "EBr3d"            
#> [163] "AME(R)"            "MB(R)"             "g1(R)"            
#> [166] "GNG"               "AL-VC4(R)"         "AL-DM2(L)"        
#> [169] "AL-DM4(L)"         "ROB(R)"            "NO(R)"            
#> [172] "PB(L7)"            "AL(R)"             "PB(L6)"           
#> [175] "AL-VA1v(R)"        "AL-VM1(R)"         "AL-VP1m(R)"       
#> [178] "CRE(-ROB,-RUB)(R)" "NO1(L)"            "PB(L1)"           
#> [181] "AL-DA3(R)"         "NO3(L)"            "AL-DP1m(R)"       
#> [184] "VMNP"              "AL-DL4(L)"         "FBl7"             
#> [187] "AL-VP1d(R)"        "b'1(R)"            "AL-DC1(L)"        
#> [190] "FB-column3"        "EBr3am"            "ATL(L)"           
#> [193] "EBr6"              "AL-VA6(L)"         "NO2(R)"           
#> [196] "a1(R)"             "AL-VL1(R)"         "b'L(R)"           
#> [199] "NO"                "OL(R)"             "AL-DM2(R)"        
#> [202] "AL-VC3m(R)"        "AL-VC5(R)"         "PENP"             
#> [205] "g2(R)"             "VES(L)"            "AL-VA1d(R)"       
#> [208] "AL-VL2a(R)"        "vACA(R)"           "hemibrain"        
#> [211] "SCL(R)"            "EBr1"              "a'2(R)"           
#> [214] "PB(L9)"            "SLP(R)"            "AL-VM5d(R)"       
#> [217] "b'2(R)"            "SIP(L)"            "PB"               
#> [220] "LX(L)"             "CAN(R)"            "LAL(R)"           
#> [223] "AL-DC3(R)"         "AL-DL2d(R)"        "LX(R)"            
#> [226] "PB(R9)"            "EBr3pw"            "LH(R)"            
#> [229] "EPA(L)"            "bL(R)"            
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "hemibrain"         "AL-VM7v(L)"        "FBl9"             
#>   [4] "CRE(-RUB)(L)"      "RUB(L)"            "g1(R)"            
#>   [7] "a'L(R)"            "CA(L)"             "ROB(R)"           
#>  [10] "AL-VA1v(R)"        "AL-VC1(R)"         "g3(R)"            
#>  [13] "NO3(L)"            "MB(+ACA)(R)"       "EPA(L)"           
#>  [16] "AL-DM6(R)"         "ATL(L)"            "gL(L)"            
#>  [19] "b1(R)"             "AL-DA1(R)"         "AL-VC3l(R)"       
#>  [22] "b'1(R)"            "AL-D(L)"           "AL-VL1(R)"        
#>  [25] "AL-VM5d(R)"        "FB"                "mALT(L)"          
#>  [28] "PB(L7)"            "FBl5"              "LAL(R)"           
#>  [31] "AL-VM4(R)"         "RUB(R)"            "lACA(R)"          
#>  [34] "PVLP(R)"           "PB(R6)"            "EBr3d"            
#>  [37] "PB(R5)"            "b2(R)"             "FLA(R)"           
#>  [40] "AL-DC4(L)"         "AL-VM7d(L)"        "PB(R3)"           
#>  [43] "INP"               "a'L(L)"            "AL-VC5(R)"        
#>  [46] "POC"               "LX(R)"             "bL(R)"            
#>  [49] "AL(R)"             "AL-DM1(L)"         "SNP(L)"           
#>  [52] "aL(L)"             "PENP"              "WED(R)"           
#>  [55] "AL-DM4(L)"         "CRE(L)"            "CA(R)"            
#>  [58] "ICL(R)"            "AL-VP1l(R)"        "PLP(R)"           
#>  [61] "LX(L)"             "AL-DC1(R)"         "AL-VA7m(R)"       
#>  [64] "AL-VC3m(R)"        "NO1(R)"            "BU(R)"            
#>  [67] "FBl1"              "mALT(R)"           "dACA(R)"          
#>  [70] "AL-DA4m(L)"        "AL-DM6(L)"         "AL-DA2(R)"        
#>  [73] "AL-DL4(R)"         "AL-VA1d(R)"        "AL-VM7v(R)"       
#>  [76] "EPA(R)"            "CAN(R)"            "NO(L)"            
#>  [79] "OL(R)"             "AMMC"              "NO"               
#>  [82] "AL-DP1l(R)"        "IB"                "CX"               
#>  [85] "GC"                "PB(L8)"            "AL-DL4(L)"        
#>  [88] "PB(R9)"            "AL-VP3(R)"         "SMP(R)"           
#>  [91] "a1(R)"             "FB-column3"        "a2(R)"            
#>  [94] "EBr2r4"            "b'L(L)"            "AL-VM1(R)"        
#>  [97] "PB(L9)"            "MB(R)"             "GNG"              
#> [100] "vACA(R)"           "AOTU(R)"           "VMNP"             
#> [103] "AL-DP1m(L)"        "FBl7"              "AL-DA4l(R)"       
#> [106] "AL-DL1(R)"         "AL-VA3(R)"         "b'2(R)"           
#> [109] "g2(R)"             "AL-VA6(L)"         "FBl8"             
#> [112] "GOR(R)"            "AL-DL2v(R)"        "AL-DM2(R)"        
#> [115] "AL-VL2a(R)"        "a'3(R)"            "SAD"              
#> [118] "ICL(L)"            "AL-DC2(R)"         "aL(R)"            
#> [121] "PB(L5)"            "PRW"               "EBr5"             
#> [124] "AL-DM5(L)"         "NO2(R)"            "a3(R)"            
#> [127] "AL-VA7l(R)"        "AL-V(R)"           "EBr3pw"           
#> [130] "VES(R)"            "AB(R)"             "EB"               
#> [133] "AL-VP1m(R)"        "PB(L2)"            "SMP(L)"           
#> [136] "NO1(L)"            "AL-DC3(R)"         "AL-DL5(R)"        
#> [139] "ME(R)"             "SAD(-AMMC)"        "PB(L1)"           
#> [142] "AL-VL2p(R)"        "AL-VM3(R)"         "AL-D(R)"          
#> [145] "AL-VA5(R)"         "AL-VM5v(R)"        "LAL(L)"           
#> [148] "GOR(L)"            "SCL(R)"            "FBl6"             
#> [151] "SLP(R)"            "AL-DM5(R)"         "AL-VP2(R)"        
#> [154] "PB(L3)"            "PB(R7)"            "MB(L)"            
#> [157] "NO(R)"             "PB(L4)"            "AL-VC2(R)"        
#> [160] "PB"                "AOT(R)"            "PED(R)"           
#> [163] "PB(R1)"            "AL-DL2d(R)"        "AB(L)"            
#> [166] "GA(R)"             "AL-DA3(R)"         "LAL(-GA)(R)"      
#> [169] "AL-VM2(R)"         "BU(L)"             "g4(R)"            
#> [172] "AL(L)"             "a'1(R)"            "g5(R)"            
#> [175] "PB(R4)"            "EBr6"              "PB(L6)"           
#> [178] "AL-DC4(R)"         "bL(L)"             "SPS(L)"           
#> [181] "AL-DA2(L)"         "FBl2"              "NO3(R)"           
#> [184] "AL-DM3(L)"         "AL-VA4(R)"         "NO2(L)"           
#> [187] "b'L(R)"            "AL-DA3(L)"         "AVLP(R)"          
#> [190] "IPS(R)"            "AL-DM1(R)"         "SPS(R)"           
#> [193] "FBl4"              "AL-DP1m(R)"        "AL-VC4(R)"        
#> [196] "AL-VP5(R)"         "AL-DC1(L)"         "AL-DM2(L)"        
#> [199] "FBl3"              "AL-DL3(R)"         "AL-DM4(R)"        
#> [202] "AL-VA6(R)"         "EBr3am"            "SNP(R)"           
#> [205] "LO(R)"             "AL-DA4m(R)"        "AL-VM7d(R)"       
#> [208] "AL-VP1d(R)"        "GF(R)"             "CRE(R)"           
#> [211] "a'2(R)"            "AL-DL5(L)"         "ATL(R)"           
#> [214] "AL-DM3(R)"         "AL-VP4(R)"         "EBr1"             
#> [217] "PB(R8)"            "VES(L)"            "LOP(R)"           
#> [220] "AL-DC2(L)"         "PB(R2)"            "AME(R)"           
#> [223] "SIP(L)"            "AL-VA2(R)"         "VLNP(R)"          
#> [226] "SCL(L)"            "SIP(R)"            "gL(R)"            
#> [229] "LH(R)"             "CRE(-ROB,-RUB)(R)"
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                      
#> [1,] "POST" "/api/npexplorer/simpleconnections"       
#> [2,] "POST" "/api/v:ver/npexplorer/autapses"          
#> [3,] "GET"  "/profile"                                
#> [4,] "GET"  "/api/npexplorer/neuronmetavals"          
#> [5,] "POST" "/api/v:ver/npexplorer/commonconnectivity"
#> [6,] "GET"  "/api/available"                          
#>      name                                                                                             
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [3,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                         
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMetaVals-fm"   
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.SetupRoutes.func1"                              

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }