Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "b2(R)"             "CA(R)"             "AL-DM2(R)"        
#>   [4] "CA(L)"             "PB(R9)"            "a'1(R)"           
#>   [7] "FBl7"              "AL-VL2a(R)"        "CRE(-ROB,-RUB)(R)"
#>  [10] "AL(L)"             "LAL(-GA)(R)"       "AL-VM7v(L)"       
#>  [13] "IPS(R)"            "AL-DC1(R)"         "AL-DL2d(R)"       
#>  [16] "AL-VM7v(R)"        "EBr5"              "a'2(R)"           
#>  [19] "CAN(R)"            "a3(R)"             "AL-DM1(R)"        
#>  [22] "AL-VM1(R)"         "AL-VP1d(R)"        "IB"               
#>  [25] "SAD"               "FLA(R)"            "AL-D(L)"          
#>  [28] "AL-DM6(L)"         "g5(R)"             "VES(L)"           
#>  [31] "a1(R)"             "b'2(R)"            "aL(R)"            
#>  [34] "CRE(L)"            "PB(R3)"            "LX(L)"            
#>  [37] "AL-DL4(R)"         "PB(L5)"            "EBr3d"            
#>  [40] "PB(L9)"            "PB(R7)"            "CX"               
#>  [43] "POC"               "CRE(R)"            "AL-DL2v(R)"       
#>  [46] "PLP(R)"            "PB(R2)"            "AL-VA1d(R)"       
#>  [49] "SPS(R)"            "b'L(R)"            "AL-DL4(L)"        
#>  [52] "FBl3"              "EPA(L)"            "AL-DM3(R)"        
#>  [55] "AL-VM3(R)"         "ATL(L)"            "EBr1"             
#>  [58] "AL-DC2(L)"         "PVLP(R)"           "AL-DC4(R)"        
#>  [61] "AL-DP1m(R)"        "SPS(L)"            "EPA(R)"           
#>  [64] "AL(R)"             "AL-VA7l(R)"        "AL-DM3(L)"        
#>  [67] "ICL(L)"            "AL-DM4(R)"         "AL-DM6(R)"        
#>  [70] "PB(R1)"            "bL(R)"             "NO(R)"            
#>  [73] "RUB(R)"            "PRW"               "g1(R)"            
#>  [76] "PB(L6)"            "LO(R)"             "ICL(R)"           
#>  [79] "AL-VA5(R)"         "AL-VC3m(R)"        "AL-V(R)"          
#>  [82] "GOR(L)"            "mALT(R)"           "AL-VP5(R)"        
#>  [85] "vACA(R)"           "AL-DA2(L)"         "OL(R)"            
#>  [88] "VES(R)"            "MB(+ACA)(R)"       "FBl8"             
#>  [91] "PB(R4)"            "AL-DL5(R)"         "AL-VM4(R)"        
#>  [94] "AL-DM4(L)"         "SAD(-AMMC)"        "AMMC"             
#>  [97] "PB(R5)"            "a'3(R)"            "AL-VA3(R)"        
#> [100] "AL-VA6(R)"         "AL-VC1(R)"         "AL-DA4m(L)"       
#> [103] "AL-DA1(R)"         "AL-VM5v(R)"        "SIP(R)"           
#> [106] "AL-DA2(R)"         "AL-VP3(R)"         "b'1(R)"           
#> [109] "SCL(R)"            "AL-VM7d(L)"        "AL-DC4(L)"        
#> [112] "LAL(R)"            "AL-DL1(R)"         "AL-VP1l(R)"       
#> [115] "PB(R6)"            "NO1(L)"            "IB112"            
#> [118] "a'L(L)"            "AL-VC3l(R)"        "SCL(L)"           
#> [121] "SMP(L)"            "PB(L7)"            "b1(R)"            
#> [124] "AL-DC2(R)"         "g3(R)"             "GA(R)"            
#> [127] "AL-DA3(L)"         "PED(R)"            "VMNP"             
#> [130] "AL-VC4(R)"         "AOT(R)"            "PB(L3)"           
#> [133] "aL(L)"             "NO3(L)"            "NO2(R)"           
#> [136] "VLNP(R)"           "lACA(R)"           "bL(L)"            
#> [139] "EB"                "RUB(L)"            "FBl2"             
#> [142] "PB(L2)"            "WED(R)"            "EBr2r4"           
#> [145] "BU(L)"             "NO(L)"             "a2(R)"            
#> [148] "AL-DM2(L)"         "PENP"              "GNG"              
#> [151] "PB(L8)"            "EBr3pw"            "dACA(R)"          
#> [154] "AL-DC1(L)"         "MB(L)"             "PB(R8)"           
#> [157] "AL-D(R)"           "AL-VP1m(R)"        "SNP(R)"           
#> [160] "gL(R)"             "AL-VL2p(R)"        "AL-VP2(R)"        
#> [163] "ROB(R)"            "AL-DA4l(R)"        "INP"              
#> [166] "AL-VC5(R)"         "FBl1"              "FBl9"             
#> [169] "AL-DP1l(R)"        "AL-DM5(L)"         "SLP(R)"           
#> [172] "AB(R)"             "SMP(R)"            "SNP(L)"           
#> [175] "FBl6"              "gL(L)"             "GOR(R)"           
#> [178] "LH(R)"             "AL-DA3(R)"         "AL-VL1(R)"        
#> [181] "AOTU(R)"           "AL-DP1m(L)"        "AL-VC2(R)"        
#> [184] "AL-DL3(R)"         "AL-VA7m(R)"        "LX(R)"            
#> [187] "LOP(R)"            "AL-DC3(R)"         "BU(R)"            
#> [190] "FBl4"              "PB(L1)"            "MB(R)"            
#> [193] "AL-VM7d(R)"        "EBr3am"            "NO2(L)"           
#> [196] "SIP(L)"            "AL-DA4m(R)"        "AL-VA4(R)"        
#> [199] "AL-VM5d(R)"        "LAL(L)"            "hemibrain"        
#> [202] "NO"                "AVLP(R)"           "PB(L4)"           
#> [205] "g4(R)"             "AL-VA6(L)"         "CRE(-RUB)(L)"     
#> [208] "AL-DL5(L)"         "AL-DM1(L)"         "FB-column3"       
#> [211] "GF(R)"             "a'L(R)"            "ATL(R)"           
#> [214] "AL-VA1v(R)"        "b'L(L)"            "GC"               
#> [217] "FB"                "AME(R)"            "AL-VP4(R)"        
#> [220] "mALT(L)"           "NO3(R)"            "ME(R)"            
#> [223] "FBl5"              "AL-DM5(R)"         "AL-VM2(R)"        
#> [226] "AB(L)"             "EBr6"              "NO1(R)"           
#> [229] "AL-VA2(R)"         "g2(R)"             "PB"               
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "b'2(R)"            "NO(L)"             "VLNP(R)"          
#>   [4] "SMP(L)"            "GC"                "NO3(R)"           
#>   [7] "SPS(L)"            "PLP(R)"            "CA(R)"            
#>  [10] "GOR(L)"            "PB"                "BU(R)"            
#>  [13] "PB(L6)"            "LAL(R)"            "a'L(L)"           
#>  [16] "OL(R)"             "CX"                "a'1(R)"           
#>  [19] "lACA(R)"           "VMNP"              "PB(R1)"           
#>  [22] "ICL(R)"            "LX(L)"             "PB(R7)"           
#>  [25] "mALT(R)"           "FB"                "ROB(R)"           
#>  [28] "VES(R)"            "b'1(R)"            "NO"               
#>  [31] "b1(R)"             "EBr5"              "PB(L7)"           
#>  [34] "LH(R)"             "PB(R2)"            "ATL(R)"           
#>  [37] "AMMC"              "hemibrain"         "SNP(R)"           
#>  [40] "EBr3am"            "EBr1"              "EB"               
#>  [43] "AOT(R)"            "RUB(R)"            "PB(L2)"           
#>  [46] "AOTU(R)"           "g2(R)"             "PB(L9)"           
#>  [49] "FBl3"              "MB(R)"             "PB(R6)"           
#>  [52] "SIP(L)"            "NO3(L)"            "SNP(L)"           
#>  [55] "FBl4"              "BU(L)"             "g4(R)"            
#>  [58] "MB(+ACA)(R)"       "FBl5"              "PB(R3)"           
#>  [61] "CAN(R)"            "AB(L)"             "SCL(L)"           
#>  [64] "CRE(-ROB,-RUB)(R)" "ATL(L)"            "PB(R8)"           
#>  [67] "GOR(R)"            "VES(L)"            "SLP(R)"           
#>  [70] "POC"               "SAD"               "NO1(R)"           
#>  [73] "EBr3d"             "EPA(L)"            "ME(R)"            
#>  [76] "gL(L)"             "FBl8"              "g3(R)"            
#>  [79] "MB(L)"             "PB(R4)"            "b'L(L)"           
#>  [82] "PB(L4)"            "a'3(R)"            "mALT(L)"          
#>  [85] "EBr6"              "dACA(R)"           "PB(L1)"           
#>  [88] "PB(L8)"            "gL(R)"             "CRE(L)"           
#>  [91] "SAD(-AMMC)"        "PVLP(R)"           "AVLP(R)"          
#>  [94] "INP"               "vACA(R)"           "LAL(L)"           
#>  [97] "PB(R5)"            "LOP(R)"            "PB(L3)"           
#> [100] "FLA(R)"            "LX(R)"             "EPA(R)"           
#> [103] "SPS(R)"            "WED(R)"            "aL(R)"            
#> [106] "CRE(R)"            "SCL(R)"            "FBl1"             
#> [109] "NO2(R)"            "AL(R)"             "SMP(R)"           
#> [112] "a'L(R)"            "FBl6"              "AME(R)"           
#> [115] "IB"                "NO1(L)"            "b2(R)"            
#> [118] "b'L(R)"            "FBl2"              "FBl7"             
#> [121] "RUB(L)"            "bL(R)"             "a3(R)"            
#> [124] "PB(L5)"            "CA(L)"             "a2(R)"            
#> [127] "LAL(-GA)(R)"       "bL(L)"             "EBr3pw"           
#> [130] "FBl9"              "PENP"              "GF(R)"            
#> [133] "PRW"               "g1(R)"             "ICL(L)"           
#> [136] "GNG"               "AL(L)"             "PB(R9)"           
#> [139] "NO(R)"             "LO(R)"             "IPS(R)"           
#> [142] "EBr2r4"            "g5(R)"             "AB(R)"            
#> [145] "NO2(L)"            "GA(R)"             "PED(R)"           
#> [148] "a'2(R)"            "SIP(R)"            "aL(L)"            
#> [151] "a1(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-VM7v(L)"        "EBr5"              "FBl4"             
#>   [4] "NO1(R)"            "a'L(R)"            "AVLP(R)"          
#>   [7] "NO1(L)"            "AL-DP1m(L)"        "AL-VA7l(R)"       
#>  [10] "AL-VM4(R)"         "AL-VP5(R)"         "SMP(L)"           
#>  [13] "EBr3pw"            "SIP(R)"            "lACA(R)"          
#>  [16] "AL-DC1(L)"         "SLP(R)"            "AL-DC2(R)"        
#>  [19] "AL-VP1l(R)"        "AL(L)"             "PB(R9)"           
#>  [22] "AL-DA4l(R)"        "NO2(L)"            "b2(R)"            
#>  [25] "BU(R)"             "AL-DA3(L)"         "AL-VA1d(R)"       
#>  [28] "ME(R)"             "AL-DM5(L)"         "PB(L7)"           
#>  [31] "g5(R)"             "GOR(R)"            "OL(R)"            
#>  [34] "AB(L)"             "LOP(R)"            "CRE(-RUB)(L)"     
#>  [37] "NO3(R)"            "a3(R)"             "PB(R5)"           
#>  [40] "AOTU(R)"           "AL-VP4(R)"         "AOT(R)"           
#>  [43] "ATL(L)"            "AL-DA1(R)"         "mALT(R)"          
#>  [46] "LH(R)"             "AL(R)"             "AL-DL2v(R)"       
#>  [49] "PED(R)"            "MB(L)"             "PB(L9)"           
#>  [52] "PB(L3)"            "AL-VC1(R)"         "AL-VC5(R)"        
#>  [55] "a1(R)"             "AL-DM5(R)"         "AL-DM6(R)"        
#>  [58] "g2(R)"             "AL-DL3(R)"         "bL(L)"            
#>  [61] "AL-DM4(L)"         "FBl9"              "AL-DL2d(R)"       
#>  [64] "AL-VL2p(R)"        "g1(R)"             "AL-DL4(L)"        
#>  [67] "PB(R2)"            "AL-DA4m(L)"        "SCL(R)"           
#>  [70] "PB(R8)"            "EPA(R)"            "CRE(R)"           
#>  [73] "SPS(R)"            "PVLP(R)"           "FBl1"             
#>  [76] "FBl3"              "bL(R)"             "AL-DA2(R)"        
#>  [79] "SIP(L)"            "PB(L4)"            "AL-DL1(R)"        
#>  [82] "SMP(R)"            "PB(L2)"            "SNP(R)"           
#>  [85] "AL-VA6(R)"         "AL-DC2(L)"         "EPA(L)"           
#>  [88] "AL-DM2(R)"         "AL-VC3l(R)"        "AB(R)"            
#>  [91] "EBr1"              "CA(R)"             "LAL(-GA)(R)"      
#>  [94] "LO(R)"             "AL-VP1d(R)"        "LAL(L)"           
#>  [97] "SPS(L)"            "FBl7"              "AL-DM3(R)"        
#> [100] "AL-VA4(R)"         "AL-VA5(R)"         "AL-VP2(R)"        
#> [103] "aL(R)"             "AL-DC4(L)"         "AL-VM1(R)"        
#> [106] "AL-VA3(R)"         "AL-VA6(L)"         "EB"               
#> [109] "AL-DM1(L)"         "a'1(R)"            "AL-DC3(R)"        
#> [112] "BU(L)"             "AL-D(R)"           "AL-VL2a(R)"       
#> [115] "gL(R)"             "AL-VM5v(R)"        "SCL(L)"           
#> [118] "PB(R1)"            "AL-VC4(R)"         "PB(L6)"           
#> [121] "AME(R)"            "NO(L)"             "AL-DL4(R)"        
#> [124] "vACA(R)"           "IPS(R)"            "GF(R)"            
#> [127] "FLA(R)"            "VES(R)"            "PB(R4)"           
#> [130] "dACA(R)"           "PB(R7)"            "AL-DM2(L)"        
#> [133] "LAL(R)"            "AL-VM3(R)"         "AL-VP1m(R)"       
#> [136] "g3(R)"             "PB(L5)"            "PB"               
#> [139] "MB(R)"             "AL-VM7d(R)"        "POC"              
#> [142] "PB(L8)"            "AL-DA4m(R)"        "AL-DC4(R)"        
#> [145] "EBr3am"            "a'2(R)"            "LX(R)"            
#> [148] "FBl5"              "FBl6"              "AL-VA1v(R)"       
#> [151] "b'1(R)"            "CA(L)"             "SAD"              
#> [154] "PRW"               "AL-DM3(L)"         "NO"               
#> [157] "ROB(R)"            "ICL(R)"            "b'L(L)"           
#> [160] "mALT(L)"           "GA(R)"             "EBr3d"            
#> [163] "FBl2"              "VLNP(R)"           "hemibrain"        
#> [166] "MB(+ACA)(R)"       "NO2(R)"            "GNG"              
#> [169] "a'L(L)"            "b'L(R)"            "AMMC"             
#> [172] "CAN(R)"            "WED(R)"            "AL-DP1l(R)"       
#> [175] "a2(R)"             "PB(L1)"            "b1(R)"            
#> [178] "AL-DL5(R)"         "AL-DM1(R)"         "AL-VM7v(R)"       
#> [181] "GOR(L)"            "NO(R)"             "CX"               
#> [184] "AL-D(L)"           "gL(L)"             "AL-DL5(L)"        
#> [187] "SNP(L)"            "INP"               "AL-VA7m(R)"       
#> [190] "PLP(R)"            "FB-column3"        "AL-VC2(R)"        
#> [193] "AL-VM2(R)"         "AL-V(R)"           "IB"               
#> [196] "AL-DM6(L)"         "VMNP"              "CRE(L)"           
#> [199] "FBl8"              "EBr6"              "CRE(-ROB,-RUB)(R)"
#> [202] "AL-DC1(R)"         "AL-DM4(R)"         "VES(L)"           
#> [205] "FB"                "AL-VA2(R)"         "AL-VL1(R)"        
#> [208] "b'2(R)"            "EBr2r4"            "g4(R)"            
#> [211] "NO3(L)"            "RUB(R)"            "GC"               
#> [214] "ICL(L)"            "ATL(R)"            "aL(L)"            
#> [217] "RUB(L)"            "AL-VM5d(R)"        "AL-VP3(R)"        
#> [220] "AL-DA2(L)"         "AL-VC3m(R)"        "a'3(R)"           
#> [223] "PENP"              "SAD(-AMMC)"        "PB(R3)"           
#> [226] "LX(L)"             "AL-DA3(R)"         "PB(R6)"           
#> [229] "AL-DP1m(R)"        "AL-VM7d(L)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "a'L(L)"            "AL-VA1v(R)"        "AL-VC3m(R)"       
#>   [4] "b'L(R)"            "CRE(-RUB)(L)"      "ATL(R)"           
#>   [7] "AL-DL2d(R)"        "AL-DP1l(R)"        "AL-VA2(R)"        
#>  [10] "EBr3am"            "ATL(L)"            "ROB(R)"           
#>  [13] "FBl7"              "AL-D(R)"           "AL-VC1(R)"        
#>  [16] "EBr3d"             "AL-DA4l(R)"        "AL-VP4(R)"        
#>  [19] "AB(L)"             "AL-DA3(L)"         "PB(R2)"           
#>  [22] "CRE(R)"            "AMMC"              "AL-DM5(R)"        
#>  [25] "PB(L8)"            "FBl9"              "CRE(L)"           
#>  [28] "CAN(R)"            "AL-DC1(R)"         "PB(R4)"           
#>  [31] "AL-VL1(R)"         "g5(R)"             "AL-VM4(R)"        
#>  [34] "LX(L)"             "AL-DP1m(R)"        "AL-VA4(R)"        
#>  [37] "AL-DL4(L)"         "LO(R)"             "CX"               
#>  [40] "g4(R)"             "AL-VM5d(R)"        "NO2(R)"           
#>  [43] "AL-VC5(R)"         "EBr2r4"            "AL-DM5(L)"        
#>  [46] "AL-VP1l(R)"        "g3(R)"             "LH(R)"            
#>  [49] "AL-DM1(R)"         "AL-DM2(R)"         "PED(R)"           
#>  [52] "SNP(L)"            "AL-VM5v(R)"        "NO(R)"            
#>  [55] "SLP(R)"            "a'L(R)"            "PB(L7)"           
#>  [58] "OL(R)"             "AL-DC4(R)"         "ME(R)"            
#>  [61] "SNP(R)"            "GOR(R)"            "NO(L)"            
#>  [64] "SMP(L)"            "NO"                "AL-VM7d(R)"       
#>  [67] "b'1(R)"            "AL-DM1(L)"         "AME(R)"           
#>  [70] "LX(R)"             "AL-VA6(L)"         "a'1(R)"           
#>  [73] "AL-VM2(R)"         "GA(R)"             "NO1(R)"           
#>  [76] "PB(L9)"            "FLA(R)"            "PB(R9)"           
#>  [79] "SAD(-AMMC)"        "VES(R)"            "FB-column3"       
#>  [82] "AB(R)"             "GF(R)"             "WED(R)"           
#>  [85] "AL-DL3(R)"         "POC"               "PRW"              
#>  [88] "FBl1"              "AL-DL4(R)"         "AL-VA1d(R)"       
#>  [91] "AL-VC2(R)"         "VMNP"              "AL-D(L)"          
#>  [94] "b2(R)"             "hemibrain"         "GC"               
#>  [97] "RUB(L)"            "AL-VA5(R)"         "PLP(R)"           
#> [100] "PB(R5)"            "EBr6"              "BU(L)"            
#> [103] "EBr1"              "AL(R)"             "PB(L3)"           
#> [106] "aL(R)"             "aL(L)"             "AL-DL5(R)"        
#> [109] "AL-VL2a(R)"        "SAD"               "PB(L1)"           
#> [112] "AL-DM6(R)"         "AL-VC4(R)"         "AL-DA2(L)"        
#> [115] "EBr3pw"            "g2(R)"             "dACA(R)"          
#> [118] "a1(R)"             "PB(R1)"            "AL-VM1(R)"        
#> [121] "AL-DC2(L)"         "a'3(R)"            "lACA(R)"          
#> [124] "EBr5"              "a3(R)"             "FBl6"             
#> [127] "AOT(R)"            "b1(R)"             "GNG"              
#> [130] "EPA(L)"            "VES(L)"            "PVLP(R)"          
#> [133] "mALT(R)"           "AL-DM4(L)"         "a'2(R)"           
#> [136] "NO2(L)"            "CA(R)"             "IPS(R)"           
#> [139] "PB(R3)"            "AL-VA7l(R)"        "AL-VL2p(R)"       
#> [142] "AL-VP5(R)"         "vACA(R)"           "PB(R8)"           
#> [145] "EB"                "FBl4"              "AL-DL2v(R)"       
#> [148] "AL-DA4m(L)"        "AL-DM3(L)"         "AL-DC1(L)"        
#> [151] "bL(L)"             "NO3(L)"            "AL-VM3(R)"        
#> [154] "AL-VP3(R)"         "IB"                "NO1(L)"           
#> [157] "EPA(R)"            "AVLP(R)"           "SCL(L)"           
#> [160] "FB"                "AL-VM7v(L)"        "AL-VM7v(R)"       
#> [163] "SPS(L)"            "AL-DM2(L)"         "FBl2"             
#> [166] "LOP(R)"            "AL-DL5(L)"         "AL-VM7d(L)"       
#> [169] "PB(L6)"            "PENP"              "AOTU(R)"          
#> [172] "bL(R)"             "AL-VA6(R)"         "PB(L5)"           
#> [175] "MB(L)"             "FBl5"              "b'L(L)"           
#> [178] "SIP(L)"            "SCL(R)"            "AL-DM4(R)"        
#> [181] "VLNP(R)"           "CRE(-ROB,-RUB)(R)" "CA(L)"            
#> [184] "gL(R)"             "MB(+ACA)(R)"       "gL(L)"            
#> [187] "FBl8"              "AL-DA4m(R)"        "RUB(R)"           
#> [190] "PB(R6)"            "a2(R)"             "ICL(L)"           
#> [193] "INP"               "LAL(R)"            "AL-DA3(R)"        
#> [196] "AL-VP1m(R)"        "AL-DM6(L)"         "AL-DP1m(L)"       
#> [199] "AL-DL1(R)"         "AL-DA1(R)"         "AL-DC2(R)"        
#> [202] "AL-DC4(L)"         "AL-VP2(R)"         "BU(R)"            
#> [205] "ICL(R)"            "NO3(R)"            "AL-DC3(R)"        
#> [208] "AL(L)"             "PB(L4)"            "SMP(R)"           
#> [211] "PB(L2)"            "LAL(L)"            "PB(R7)"           
#> [214] "AL-VC3l(R)"        "SIP(R)"            "AL-DM3(R)"        
#> [217] "AL-V(R)"           "LAL(-GA)(R)"       "AL-VA7m(R)"       
#> [220] "FBl3"              "AL-VA3(R)"         "SPS(R)"           
#> [223] "PB"                "GOR(L)"            "MB(R)"            
#> [226] "AL-DA2(R)"         "mALT(L)"           "g1(R)"            
#> [229] "AL-VP1d(R)"        "b'2(R)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method     path                                   
#> [1,] "GET"      "/api/v:ver/npexplorer/rankedtable"    
#> [2,] "GET"      "/api/v:ver/dbmeta/instances"          
#> [3,] "HEAD"     "/api"                                 
#> [4,] "PROPFIND" "/api/*"                               
#> [5,] "GET"      "/api/v:ver/npexplorer/roiconnectivity"
#> [6,] "POST"     "/api/npexplorer/roiconnectivity"      
#>      name                                                                                           
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [3,] "github.com/labstack/echo.glob..func1"                                                         
#> [4,] "github.com/labstack/echo.glob..func1"                                                         
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIConnectivity-fm"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }