Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "EB"                "FBl8"              "LH(R)"            
#>   [4] "ATL(R)"            "bL(R)"             "VES(L)"           
#>   [7] "PB(L5)"            "b'2(R)"            "a2(R)"            
#>  [10] "SMP(L)"            "SIP(L)"            "LAL(-GA)(R)"      
#>  [13] "a'L(R)"            "PB(L6)"            "LAL(R)"           
#>  [16] "CA(L)"             "AOTU(R)"           "g4(R)"            
#>  [19] "PVLP(R)"           "FBl9"              "SLP(R)"           
#>  [22] "PENP"              "WED(R)"            "SCL(R)"           
#>  [25] "FBl2"              "AVLP(R)"           "AB(R)"            
#>  [28] "IPS(R)"            "a'L(L)"            "PB(L3)"           
#>  [31] "EBr3pw"            "ICL(L)"            "AOT(R)"           
#>  [34] "BU(L)"             "FLA(R)"            "PLP(R)"           
#>  [37] "EPA(R)"            "PB(L9)"            "vACA(R)"          
#>  [40] "NO2(L)"            "PB(L1)"            "PB(R8)"           
#>  [43] "PB"                "EBr2r4"            "NO(R)"            
#>  [46] "PB(L7)"            "PB(R3)"            "PB(R7)"           
#>  [49] "PED(R)"            "FB"                "PB(L2)"           
#>  [52] "CRE(-ROB,-RUB)(R)" "LAL(L)"            "GF(R)"            
#>  [55] "GA(R)"             "POC"               "hemibrain"        
#>  [58] "SPS(L)"            "BU(R)"             "SAD(-AMMC)"       
#>  [61] "FBl6"              "b1(R)"             "CRE(R)"           
#>  [64] "LX(R)"             "gL(L)"             "PB(R4)"           
#>  [67] "AL(R)"             "PB(R9)"            "NO2(R)"           
#>  [70] "EPA(L)"            "SCL(L)"            "SPS(R)"           
#>  [73] "g2(R)"             "ROB(R)"            "GOR(R)"           
#>  [76] "a'3(R)"            "AMMC"              "lACA(R)"          
#>  [79] "NO3(L)"            "ICL(R)"            "g3(R)"            
#>  [82] "GOR(L)"            "ME(R)"             "gL(R)"            
#>  [85] "LO(R)"             "VES(R)"            "FBl4"             
#>  [88] "NO3(R)"            "a'1(R)"            "LX(L)"            
#>  [91] "INP"               "OL(R)"             "PB(R5)"           
#>  [94] "ATL(L)"            "b'L(R)"            "EBr5"             
#>  [97] "AL(L)"             "FBl5"              "PB(L4)"           
#> [100] "VLNP(R)"           "mALT(L)"           "PRW"              
#> [103] "GC"                "FBl3"              "PB(R2)"           
#> [106] "aL(L)"             "AB(L)"             "NO1(R)"           
#> [109] "SIP(R)"            "FBl1"              "RUB(L)"           
#> [112] "RUB(R)"            "LOP(R)"            "b2(R)"            
#> [115] "MB(L)"             "bL(L)"             "EBr3am"           
#> [118] "aL(R)"             "EBr3d"             "SNP(R)"           
#> [121] "NO"                "CRE(L)"            "GNG"              
#> [124] "CX"                "mALT(R)"           "NO1(L)"           
#> [127] "PB(R1)"            "SNP(L)"            "CA(R)"            
#> [130] "MB(R)"             "a1(R)"             "a3(R)"            
#> [133] "PB(R6)"            "SAD"               "dACA(R)"          
#> [136] "a'2(R)"            "MB(+ACA)(R)"       "FBl7"             
#> [139] "CAN(R)"            "b'L(L)"            "IB"               
#> [142] "EBr6"              "SMP(R)"            "g5(R)"            
#> [145] "g1(R)"             "EBr1"              "PB(L8)"           
#> [148] "VMNP"              "AME(R)"            "NO(L)"            
#> [151] "b'1(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                                     
#> [1,] "TRACE"   "/api"                                   
#> [2,] "GET"     "/api/v:ver/npexplorer/distribution"     
#> [3,] "POST"    "/api/v:ver/npexplorer/completeness"     
#> [4,] "GET"     "/api/dbmeta/datasets"                   
#> [5,] "OPTIONS" "/api"                                   
#> [6,] "POST"    "/api/v:ver/roimeshes/mesh/:dataset/:roi"
#>      name                                                                            
#> [1,] "github.com/labstack/echo.glob..func1"                                          
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"            
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"            
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatasets-fm"
#> [5,] "github.com/labstack/echo.glob..func1"                                          
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"            

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }