Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-VM7v(L)"        "PB(L6)"            "AL-VC1(R)"        
#>   [4] "AL-VC2(R)"         "AL-VM1(R)"         "EBr1"             
#>   [7] "EBr6"              "FB-column3"        "GNG"              
#>  [10] "a3(R)"             "NO3(R)"            "AL-VM2(R)"        
#>  [13] "INP"               "AL-DL2v(R)"        "NO3(L)"           
#>  [16] "EPA(L)"            "a'1(R)"            "PVLP(R)"          
#>  [19] "AL-DM4(R)"         "WED(R)"            "SPS(R)"           
#>  [22] "AL-DL2d(R)"        "SAD"               "SNP(L)"           
#>  [25] "AL-DA2(R)"         "AL-VA6(R)"         "AL-DA3(L)"        
#>  [28] "a'L(R)"            "LH(R)"             "a'2(R)"           
#>  [31] "gL(L)"             "NO"                "PB(R3)"           
#>  [34] "a1(R)"             "BU(L)"             "hemibrain"        
#>  [37] "SLP(R)"            "AL-VM7d(R)"        "SCL(L)"           
#>  [40] "g3(R)"             "PB(L5)"            "ICL(L)"           
#>  [43] "AL-DL1(R)"         "SPS(L)"            "aL(R)"            
#>  [46] "PB(R2)"            "AME(R)"            "vACA(R)"          
#>  [49] "GC"                "AL-DP1m(R)"        "AL-VC3l(R)"       
#>  [52] "IPS(R)"            "AL-DM2(R)"         "AL-VA2(R)"        
#>  [55] "AL-VA3(R)"         "AL-VL2p(R)"        "AL-DA4m(R)"       
#>  [58] "g5(R)"             "AL-VP5(R)"         "VLNP(R)"          
#>  [61] "b'2(R)"            "AL-DP1m(L)"        "AL(R)"            
#>  [64] "AL-VA7m(R)"        "AL-VP3(R)"         "a2(R)"            
#>  [67] "NO2(L)"            "b'1(R)"            "AL-DA4l(R)"       
#>  [70] "RUB(R)"            "lACA(R)"           "NO(L)"            
#>  [73] "AL-DM5(R)"         "AOT(R)"            "NO2(R)"           
#>  [76] "LAL(R)"            "AL-DL5(L)"         "AL-DM4(L)"        
#>  [79] "EBr3pw"            "AL-VM5v(R)"        "AL-V(R)"          
#>  [82] "GA(R)"             "CAN(R)"            "LAL(-GA)(R)"      
#>  [85] "AL-D(L)"           "AL-VA1v(R)"        "AL-VM5d(R)"       
#>  [88] "AL-VP4(R)"         "PB(L8)"            "AL-DC1(L)"        
#>  [91] "POC"               "EBr5"              "PB(L9)"           
#>  [94] "AVLP(R)"           "GOR(R)"            "AL-DP1l(R)"       
#>  [97] "AL-VM3(R)"         "CRE(R)"            "gL(R)"            
#> [100] "PB(L7)"            "FBl9"              "AL-VP1m(R)"       
#> [103] "NO(R)"             "PB(R1)"            "CRE(L)"           
#> [106] "ATL(R)"            "AL-DL5(R)"         "NO1(R)"           
#> [109] "PB(L1)"            "MB(+ACA)(R)"       "a'L(L)"           
#> [112] "SMP(L)"            "ROB(R)"            "ICL(R)"           
#> [115] "AL-DA1(R)"         "AL-VL2a(R)"        "PB(R5)"           
#> [118] "AL-VA6(L)"         "b'L(L)"            "AL-VM7v(R)"       
#> [121] "EBr3am"            "mALT(R)"           "FLA(R)"           
#> [124] "PB"                "PB(R9)"            "PB(L4)"           
#> [127] "PENP"              "AL-DL4(R)"         "AL-DL4(L)"        
#> [130] "LX(R)"             "AL-D(R)"           "AL-VL1(R)"        
#> [133] "PRW"               "bL(L)"             "GF(R)"            
#> [136] "AL-DA2(L)"         "LAL(L)"            "GOR(L)"           
#> [139] "b2(R)"             "ME(R)"             "AL-DL3(R)"        
#> [142] "PED(R)"            "FBl1"              "MB(R)"            
#> [145] "AL-DM6(R)"         "AMMC"              "VES(R)"           
#> [148] "AL-DM6(L)"         "FBl3"              "b1(R)"            
#> [151] "aL(L)"             "a'3(R)"            "SCL(R)"           
#> [154] "EPA(R)"            "FBl7"              "PLP(R)"           
#> [157] "AL-DA4m(L)"        "FBl2"              "ATL(L)"           
#> [160] "PB(R7)"            "EB"                "SIP(R)"           
#> [163] "RUB(L)"            "AL-VA5(R)"         "PB(L3)"           
#> [166] "AL-DC1(R)"         "CX"                "FB"               
#> [169] "AL-DC4(L)"         "FBl4"              "AL-DC3(R)"        
#> [172] "MB(L)"             "AL-DM5(L)"         "SAD(-AMMC)"       
#> [175] "FBl6"              "IB"                "AOTU(R)"          
#> [178] "g4(R)"             "SMP(R)"            "b'L(R)"           
#> [181] "PB(R4)"            "AL-DM1(R)"         "PB(L2)"           
#> [184] "FBl5"              "AL-VA1d(R)"        "AL-VC4(R)"        
#> [187] "AL-VC5(R)"         "bL(R)"             "AL-DC2(L)"        
#> [190] "AL-DM2(L)"         "AL-DC2(R)"         "AL-VA4(R)"        
#> [193] "AL-VM4(R)"         "SIP(L)"            "mALT(L)"          
#> [196] "AL(L)"             "IB112"             "EBr3d"            
#> [199] "EBr2r4"            "AL-VM7d(L)"        "CA(R)"            
#> [202] "OL(R)"             "AB(L)"             "g2(R)"            
#> [205] "AL-VP2(R)"         "AL-DM1(L)"         "AL-VC3m(R)"       
#> [208] "AL-VP1l(R)"        "AL-DC4(R)"         "LO(R)"            
#> [211] "LX(L)"             "CRE(-ROB,-RUB)(R)" "PB(R8)"           
#> [214] "VES(L)"            "AL-DM3(R)"         "LOP(R)"           
#> [217] "CA(L)"             "PB(R6)"            "SNP(R)"           
#> [220] "BU(R)"             "CRE(-RUB)(L)"      "FBl8"             
#> [223] "AB(R)"             "AL-VA7l(R)"        "VMNP"             
#> [226] "NO1(L)"            "dACA(R)"           "AL-DM3(L)"        
#> [229] "AL-DA3(R)"         "AL-VP1d(R)"        "g1(R)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "RUB(L)"            "b'1(R)"            "AOTU(R)"          
#>   [4] "EBr3pw"            "PB(R4)"            "a'L(L)"           
#>   [7] "PB(L3)"            "VLNP(R)"           "NO2(L)"           
#>  [10] "PB(R6)"            "AME(R)"            "PB(L4)"           
#>  [13] "ICL(L)"            "EBr1"              "NO(R)"            
#>  [16] "MB(+ACA)(R)"       "LX(L)"             "OL(R)"            
#>  [19] "VES(L)"            "SLP(R)"            "FB"               
#>  [22] "GOR(L)"            "g4(R)"             "b1(R)"            
#>  [25] "NO(L)"             "b'2(R)"            "PB(L6)"           
#>  [28] "PB(R2)"            "IPS(R)"            "AB(L)"            
#>  [31] "EBr5"              "PB(L1)"            "a'2(R)"           
#>  [34] "a'L(R)"            "SIP(R)"            "EPA(L)"           
#>  [37] "dACA(R)"           "ROB(R)"            "GF(R)"            
#>  [40] "SIP(L)"            "EB"                "PB(R9)"           
#>  [43] "SNP(L)"            "LO(R)"             "CA(R)"            
#>  [46] "AVLP(R)"           "hemibrain"         "MB(R)"            
#>  [49] "AL(R)"             "a3(R)"             "PB(R8)"           
#>  [52] "bL(L)"             "CAN(R)"            "LOP(R)"           
#>  [55] "vACA(R)"           "AMMC"              "g2(R)"            
#>  [58] "MB(L)"             "b'L(R)"            "CRE(-ROB,-RUB)(R)"
#>  [61] "PED(R)"            "bL(R)"             "IB"               
#>  [64] "PB(L9)"            "aL(R)"             "NO1(L)"           
#>  [67] "gL(L)"             "POC"               "CX"               
#>  [70] "SMP(L)"            "FBl1"              "SPS(L)"           
#>  [73] "PB(R7)"            "EBr6"              "EPA(R)"           
#>  [76] "VES(R)"            "FBl7"              "LAL(L)"           
#>  [79] "FLA(R)"            "PB"                "PB(R1)"           
#>  [82] "CRE(L)"            "PVLP(R)"           "FBl4"             
#>  [85] "INP"               "g1(R)"             "a'3(R)"           
#>  [88] "mALT(L)"           "GNG"               "gL(R)"            
#>  [91] "FBl5"              "aL(L)"             "ATL(L)"           
#>  [94] "NO"                "b2(R)"             "ME(R)"            
#>  [97] "FBl9"              "NO3(R)"            "GC"               
#> [100] "WED(R)"            "FBl6"              "a1(R)"            
#> [103] "SAD"               "SNP(R)"            "FBl8"             
#> [106] "b'L(L)"            "VMNP"              "GA(R)"            
#> [109] "AOT(R)"            "EBr3am"            "g3(R)"            
#> [112] "PB(L8)"            "NO1(R)"            "NO3(L)"           
#> [115] "g5(R)"             "FBl2"              "LH(R)"            
#> [118] "ATL(R)"            "CRE(R)"            "FBl3"             
#> [121] "a'1(R)"            "SCL(R)"            "CA(L)"            
#> [124] "lACA(R)"           "PB(L5)"            "PRW"              
#> [127] "PB(R5)"            "SAD(-AMMC)"        "LAL(-GA)(R)"      
#> [130] "LAL(R)"            "PB(L2)"            "PENP"             
#> [133] "PLP(R)"            "LX(R)"             "BU(R)"            
#> [136] "AL(L)"             "SMP(R)"            "mALT(R)"          
#> [139] "BU(L)"             "EBr3d"             "EBr2r4"           
#> [142] "PB(L7)"            "AB(R)"             "NO2(R)"           
#> [145] "GOR(R)"            "a2(R)"             "RUB(R)"           
#> [148] "PB(R3)"            "ICL(R)"            "SCL(L)"           
#> [151] "SPS(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DM4(L)"         "SIP(R)"            "NO3(L)"           
#>   [4] "AMMC"              "AL-DL2d(R)"        "LAL(L)"           
#>   [7] "SPS(R)"            "dACA(R)"           "PB(R8)"           
#>  [10] "FBl5"              "AL-DC4(L)"         "AL-VC1(R)"        
#>  [13] "mALT(L)"           "a'L(R)"            "AL-DM6(L)"        
#>  [16] "b'L(L)"            "NO1(R)"            "CRE(L)"           
#>  [19] "NO3(R)"            "BU(R)"             "PLP(R)"           
#>  [22] "CA(R)"             "AL-VM4(R)"         "PB(L5)"           
#>  [25] "GOR(L)"            "b2(R)"             "AL-DC1(L)"        
#>  [28] "AL-VM7v(L)"        "OL(R)"             "GF(R)"            
#>  [31] "MB(+ACA)(R)"       "SLP(R)"            "AL-VP1d(R)"       
#>  [34] "RUB(R)"            "AL-VM7d(L)"        "EBr3pw"           
#>  [37] "AVLP(R)"           "a3(R)"             "AL-DL1(R)"        
#>  [40] "PB(R7)"            "SNP(L)"            "AL-DM1(R)"        
#>  [43] "ATL(R)"            "AL-VM3(R)"         "AL-VM7d(R)"       
#>  [46] "AL-VP4(R)"         "PB(R1)"            "FBl7"             
#>  [49] "AME(R)"            "CAN(R)"            "a'2(R)"           
#>  [52] "AL-DA3(L)"         "FBl1"              "LAL(R)"           
#>  [55] "AL-VA1v(R)"        "AL-DA4m(L)"        "AL-DC2(L)"        
#>  [58] "AL-VA6(L)"         "FBl4"              "AL-DC1(R)"        
#>  [61] "EBr5"              "PB"                "SAD"              
#>  [64] "AL-DP1l(R)"        "AL-VA4(R)"         "mALT(R)"          
#>  [67] "GC"                "a'1(R)"            "AL-DL5(R)"        
#>  [70] "CX"                "AOTU(R)"           "aL(L)"            
#>  [73] "AL-DM2(R)"         "a'3(R)"            "AL-VC4(R)"        
#>  [76] "PB(R2)"            "AL-DA4m(R)"        "AL-DL4(R)"        
#>  [79] "PB(L2)"            "a2(R)"             "ICL(L)"           
#>  [82] "AL-V(R)"           "PVLP(R)"           "AL-DC3(R)"        
#>  [85] "AL-VM1(R)"         "PED(R)"            "FB"               
#>  [88] "EBr6"              "PB(L9)"            "bL(L)"            
#>  [91] "b'1(R)"            "VLNP(R)"           "g1(R)"            
#>  [94] "PB(R9)"            "MB(R)"             "AL-VC3m(R)"       
#>  [97] "LAL(-GA)(R)"       "FB-column3"        "AL-DM1(L)"        
#> [100] "FBl2"              "AL-DM6(R)"         "AL-VL2a(R)"       
#> [103] "NO"                "SAD(-AMMC)"        "PB(L3)"           
#> [106] "FLA(R)"            "SPS(L)"            "SNP(R)"           
#> [109] "AL-DM2(L)"         "PB(L8)"            "CRE(-RUB)(L)"     
#> [112] "AL-DM5(R)"         "AL-VA7l(R)"        "b'L(R)"           
#> [115] "GOR(R)"            "AL-VA6(R)"         "a1(R)"            
#> [118] "WED(R)"            "aL(R)"             "AL(L)"            
#> [121] "GNG"               "AL-VC3l(R)"        "AL-VC5(R)"        
#> [124] "g3(R)"             "g4(R)"             "AL-DM3(L)"        
#> [127] "NO(R)"             "ICL(R)"            "AL-DC4(R)"        
#> [130] "ATL(L)"            "AL-VL1(R)"         "AL-VP1m(R)"       
#> [133] "IB"                "PENP"              "PB(R6)"           
#> [136] "AL-D(L)"           "b1(R)"             "AL-DA3(R)"        
#> [139] "CRE(-ROB,-RUB)(R)" "VMNP"              "AL-DM3(R)"        
#> [142] "SCL(L)"            "hemibrain"         "AL-DA1(R)"        
#> [145] "PRW"               "gL(L)"             "LO(R)"            
#> [148] "IPS(R)"            "AL-VL2p(R)"        "POC"              
#> [151] "PB(R5)"            "PB(L1)"            "SMP(R)"           
#> [154] "FBl8"              "a'L(L)"            "EBr3am"           
#> [157] "AL-DL2v(R)"        "AB(L)"             "g2(R)"            
#> [160] "EB"                "AL-DP1m(L)"        "LX(L)"            
#> [163] "RUB(L)"            "ME(R)"             "INP"              
#> [166] "PB(R4)"            "AL-VP5(R)"         "AL-VP3(R)"        
#> [169] "lACA(R)"           "AB(R)"             "bL(R)"            
#> [172] "PB(L4)"            "AL-VC2(R)"         "AL-VM2(R)"        
#> [175] "VES(L)"            "b'2(R)"            "BU(L)"            
#> [178] "EBr2r4"            "AL-VM7v(R)"        "MB(L)"            
#> [181] "FBl9"              "AL-VA7m(R)"        "vACA(R)"          
#> [184] "CA(L)"             "LX(R)"             "AL(R)"            
#> [187] "AL-DL4(L)"         "EPA(L)"            "AL-DA2(R)"        
#> [190] "AL-D(R)"           "AOT(R)"            "SMP(L)"           
#> [193] "ROB(R)"            "gL(R)"             "FBl3"             
#> [196] "PB(R3)"            "AL-VA2(R)"         "AL-DL5(L)"        
#> [199] "VES(R)"            "NO2(R)"            "NO(L)"            
#> [202] "AL-DM4(R)"         "AL-DC2(R)"         "g5(R)"            
#> [205] "AL-DP1m(R)"        "AL-VA5(R)"         "AL-VP2(R)"        
#> [208] "NO2(L)"            "NO1(L)"            "CRE(R)"           
#> [211] "EPA(R)"            "AL-DM5(L)"         "FBl6"             
#> [214] "SCL(R)"            "PB(L7)"            "AL-DL3(R)"        
#> [217] "LOP(R)"            "SIP(L)"            "GA(R)"            
#> [220] "EBr3d"             "AL-DA4l(R)"        "AL-VA1d(R)"       
#> [223] "AL-VM5d(R)"        "AL-VM5v(R)"        "AL-VP1l(R)"       
#> [226] "EBr1"              "AL-DA2(L)"         "PB(L6)"           
#> [229] "LH(R)"             "AL-VA3(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "a'L(R)"            "AL(L)"             "AL-DA2(L)"        
#>   [4] "AL-VA6(L)"         "SCL(R)"            "PB(L7)"           
#>   [7] "FB-column3"        "PB(L4)"            "AL-DM4(R)"        
#>  [10] "AL-VA5(R)"         "AL-VM7d(R)"        "b'2(R)"           
#>  [13] "FBl4"              "AL-DA4m(R)"        "AL-DM1(R)"        
#>  [16] "AL-DC2(L)"         "FBl7"              "AL-DC1(L)"        
#>  [19] "AL-DM4(L)"         "PB(L6)"            "AL-VA2(R)"        
#>  [22] "AL-DA3(L)"         "VES(R)"            "EBr3am"           
#>  [25] "PLP(R)"            "a1(R)"             "LX(R)"            
#>  [28] "AL-DL5(L)"         "FBl5"              "AL-DL3(R)"        
#>  [31] "EBr6"              "AL-DL5(R)"         "AL-VM4(R)"        
#>  [34] "bL(L)"             "g3(R)"             "a3(R)"            
#>  [37] "VES(L)"            "AL-VA7m(R)"        "AL-DC2(R)"        
#>  [40] "b2(R)"             "SPS(L)"            "CRE(-RUB)(L)"     
#>  [43] "AL-VA1d(R)"        "AL-VC1(R)"         "AL-VC2(R)"        
#>  [46] "AL-VM5v(R)"        "EBr1"              "ICL(L)"           
#>  [49] "LAL(L)"            "a2(R)"             "AL-DM6(L)"        
#>  [52] "FBl2"              "FBl8"              "SIP(L)"           
#>  [55] "dACA(R)"           "PB(R9)"            "AL-VC5(R)"        
#>  [58] "VLNP(R)"           "GA(R)"             "EBr3pw"           
#>  [61] "AL-VL2p(R)"        "SCL(L)"            "CA(R)"            
#>  [64] "AL-DM3(R)"         "PB(L8)"            "a'1(R)"           
#>  [67] "LO(R)"             "EBr5"              "EPA(R)"           
#>  [70] "FBl3"              "MB(R)"             "AL-VC3m(R)"       
#>  [73] "AL-VM1(R)"         "AL-VP1l(R)"        "lACA(R)"          
#>  [76] "AL-VM7d(L)"        "FBl9"              "PB(R3)"           
#>  [79] "AL-DA2(R)"         "LOP(R)"            "CX"               
#>  [82] "NO1(L)"            "g4(R)"             "AME(R)"           
#>  [85] "a'L(L)"            "AL-DC3(R)"         "AL-VC4(R)"        
#>  [88] "RUB(R)"            "AL-DC4(L)"         "ICL(R)"           
#>  [91] "AL-VA6(R)"         "GOR(L)"            "ATL(R)"           
#>  [94] "NO(L)"             "AL-VA1v(R)"        "MB(L)"            
#>  [97] "AL-VM2(R)"         "AL-VP3(R)"         "AL-VP1m(R)"       
#> [100] "AMMC"              "BU(L)"             "AL-DM3(L)"        
#> [103] "PB(R2)"            "b1(R)"             "bL(R)"            
#> [106] "ROB(R)"            "IPS(R)"            "SLP(R)"           
#> [109] "AL-DC4(R)"         "b'L(R)"            "ME(R)"            
#> [112] "EB"                "SAD"               "AL-DA4m(L)"       
#> [115] "AL-VA4(R)"         "PB(L5)"            "mALT(L)"          
#> [118] "PRW"               "FBl1"              "RUB(L)"           
#> [121] "LAL(R)"            "AL-DL2v(R)"        "AB(L)"            
#> [124] "EBr2r4"            "b'1(R)"            "PB(R7)"           
#> [127] "NO3(R)"            "LX(L)"             "AL-VA3(R)"        
#> [130] "CRE(R)"            "AL-DA1(R)"         "AL-DM6(R)"        
#> [133] "VMNP"              "GC"                "AL-DM5(L)"        
#> [136] "PENP"              "SPS(R)"            "AL-DP1m(L)"       
#> [139] "CRE(-ROB,-RUB)(R)" "NO"                "NO2(L)"           
#> [142] "hemibrain"         "mALT(R)"           "b'L(L)"           
#> [145] "PB(R6)"            "PB(L1)"            "PB(R1)"           
#> [148] "g5(R)"             "SNP(L)"            "AVLP(R)"          
#> [151] "AB(R)"             "AL-DL4(R)"         "AL-V(R)"          
#> [154] "AL-VP1d(R)"        "aL(R)"             "NO3(L)"           
#> [157] "FBl6"              "PB(R5)"            "SMP(L)"           
#> [160] "SIP(R)"            "CRE(L)"            "AL-VP5(R)"        
#> [163] "IB"                "AL-VM7v(L)"        "PB(L9)"           
#> [166] "EPA(L)"            "FB"                "AL-DC1(R)"        
#> [169] "AL-DM2(R)"         "AL-VP2(R)"         "LH(R)"            
#> [172] "AL-DM5(R)"         "AL-DM2(L)"         "AL-DP1m(R)"       
#> [175] "AL-VM3(R)"         "SAD(-AMMC)"        "PB(R4)"           
#> [178] "OL(R)"             "PB(R8)"            "POC"              
#> [181] "PB(L2)"            "PB"                "EBr3d"            
#> [184] "AL-DA3(R)"         "AL-DL1(R)"         "PED(R)"           
#> [187] "g2(R)"             "NO2(R)"            "GOR(R)"           
#> [190] "AL-VP4(R)"         "g1(R)"             "NO1(R)"           
#> [193] "PVLP(R)"           "AL-VC3l(R)"        "BU(R)"            
#> [196] "AL(R)"             "a'3(R)"            "vACA(R)"          
#> [199] "ATL(L)"            "a'2(R)"            "AL-DL4(L)"        
#> [202] "AL-VL1(R)"         "AOT(R)"            "AOTU(R)"          
#> [205] "CAN(R)"            "GF(R)"             "SNP(R)"           
#> [208] "AL-D(L)"           "INP"               "AL-VL2a(R)"       
#> [211] "MB(+ACA)(R)"       "AL-DP1l(R)"        "CA(L)"            
#> [214] "LAL(-GA)(R)"       "gL(L)"             "NO(R)"            
#> [217] "gL(R)"             "aL(L)"             "FLA(R)"           
#> [220] "AL-DM1(L)"         "GNG"               "AL-DA4l(R)"       
#> [223] "AL-DL2d(R)"        "PB(L3)"            "WED(R)"           
#> [226] "AL-D(R)"           "AL-VA7l(R)"        "AL-VM7v(R)"       
#> [229] "SMP(R)"            "AL-VM5d(R)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                           
#> [1,] "GET"  "/api/v:ver/npexplorer/rankedtable"            
#> [2,] "POST" "/api/v:ver/npexplorer/roisinneuron"           
#> [3,] "GET"  "/api/v:ver/npexplorer/celltype/:dataset/:type"
#> [4,] "GET"  "/api/cached/roiconnectivity"                  
#> [5,] "HEAD" "/api/*"                                       
#> [6,] "POST" "/api/npexplorer/findneurons"                  
#>      name                                                                                       
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROIConnectivity-fm"   
#> [5,] "github.com/labstack/echo.glob..func1"                                                     
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getFindNeurons-fm"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }