Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "dACA(R)"           "IPS(R)"            "AL-VC3m(R)"       
#>   [4] "PB(L8)"            "AL-VM7d(L)"        "PB(L9)"           
#>   [7] "CRE(-ROB,-RUB)(R)" "FBl4"              "FBl7"             
#>  [10] "PENP"              "PB(R7)"            "AL-DM2(L)"        
#>  [13] "NO"                "gL(R)"             "FBl3"             
#>  [16] "RUB(R)"            "mALT(L)"           "a1(R)"            
#>  [19] "NO1(R)"            "g4(R)"             "bL(L)"            
#>  [22] "FBl6"              "AMMC"              "AL-DA3(L)"        
#>  [25] "LO(R)"             "AL-DM1(R)"         "hemibrain"        
#>  [28] "BU(L)"             "AL-DM3(L)"         "IB112"            
#>  [31] "AL-VP1d(R)"        "SCL(L)"            "a2(R)"            
#>  [34] "CAN(R)"            "AL-DA4m(R)"        "SPS(R)"           
#>  [37] "PB(R3)"            "ICL(R)"            "PB(L5)"           
#>  [40] "NO3(R)"            "PB(R1)"            "AL-VA2(R)"        
#>  [43] "AL-VM7v(R)"        "PB(L2)"            "AL-DM5(L)"        
#>  [46] "NO2(R)"            "AL-VP5(R)"         "AL-DM3(R)"        
#>  [49] "AL-VL2a(R)"        "AL-VP2(R)"         "b'1(R)"           
#>  [52] "AL-DC4(R)"         "CA(L)"             "SMP(R)"           
#>  [55] "mALT(R)"           "CRE(-RUB)(L)"      "GF(R)"            
#>  [58] "SMP(L)"            "AL-DL2v(R)"        "AL-VA7l(R)"       
#>  [61] "WED(R)"            "g5(R)"             "MB(+ACA)(R)"      
#>  [64] "LH(R)"             "AL-DM2(R)"         "AL-VM5v(R)"       
#>  [67] "AL-DP1m(L)"        "AL-VP4(R)"         "AL-VP1l(R)"       
#>  [70] "SCL(R)"            "EBr1"              "aL(R)"            
#>  [73] "EBr3am"            "vACA(R)"           "GC"               
#>  [76] "PB(L1)"            "AL-DL5(R)"         "AL-VA1v(R)"       
#>  [79] "AL-VC4(R)"         "PED(R)"            "AL-VA3(R)"        
#>  [82] "NO1(L)"            "AL-DL4(R)"         "FB"               
#>  [85] "AL-DL1(R)"         "lACA(R)"           "AL-VA6(L)"        
#>  [88] "AOT(R)"            "b'L(R)"            "SIP(R)"           
#>  [91] "AL-DA1(R)"         "b2(R)"             "EBr6"             
#>  [94] "VMNP"              "SAD(-AMMC)"        "FB-column3"       
#>  [97] "AL-DM4(R)"         "g3(R)"             "ROB(R)"           
#> [100] "CRE(L)"            "AVLP(R)"           "AL-VL1(R)"        
#> [103] "AL-VM7v(L)"        "g1(R)"             "b1(R)"            
#> [106] "CX"                "b'2(R)"            "g2(R)"            
#> [109] "b'L(L)"            "AL-DL3(R)"         "AL-VM1(R)"        
#> [112] "LAL(-GA)(R)"       "GOR(L)"            "AL-D(L)"          
#> [115] "OL(R)"             "AL-VM4(R)"         "AL-VM7d(R)"       
#> [118] "PLP(R)"            "AL-DA2(R)"         "VES(R)"           
#> [121] "aL(L)"             "AL-VM3(R)"         "AL-VM5d(R)"       
#> [124] "a'3(R)"            "GOR(R)"            "AL-DP1l(R)"       
#> [127] "AL-VA5(R)"         "AL-DA2(L)"         "AL-DC1(L)"        
#> [130] "LAL(R)"            "a'L(L)"            "AL-DA4l(R)"       
#> [133] "AL-VC2(R)"         "AL(R)"             "AL-DC2(R)"        
#> [136] "AL-DM4(L)"         "PB(L7)"            "FBl1"             
#> [139] "ATL(R)"            "AB(R)"             "INP"              
#> [142] "EPA(L)"            "a3(R)"             "PB(L4)"           
#> [145] "PB(R2)"            "RUB(L)"            "MB(R)"            
#> [148] "AL-DC1(R)"         "LOP(R)"            "EBr3d"            
#> [151] "FBl2"              "FBl9"              "bL(R)"            
#> [154] "AL-V(R)"           "a'L(R)"            "EPA(R)"           
#> [157] "AL-DC2(L)"         "ATL(L)"            "MB(L)"            
#> [160] "GNG"               "VES(L)"            "AL-DA3(R)"        
#> [163] "AL-VA7m(R)"        "IB"                "gL(L)"            
#> [166] "AL-DA4m(L)"        "AL-D(R)"           "VLNP(R)"          
#> [169] "NO2(L)"            "AL-VC3l(R)"        "AL-VL2p(R)"       
#> [172] "POC"               "BU(R)"             "AL(L)"            
#> [175] "AL-DM1(L)"         "PB(R9)"            "SNP(L)"           
#> [178] "AL-VA4(R)"         "FBl8"              "AL-DM6(R)"        
#> [181] "AL-VA6(R)"         "PB(R5)"            "EBr5"             
#> [184] "ME(R)"             "a'1(R)"            "NO3(L)"           
#> [187] "PB(R6)"            "LX(R)"             "AL-DM6(L)"        
#> [190] "AME(R)"            "PB(R4)"            "FLA(R)"           
#> [193] "NO(L)"             "PVLP(R)"           "AL-DL2d(R)"       
#> [196] "AL-VP1m(R)"        "AOTU(R)"           "PB(L6)"           
#> [199] "PB"                "EBr3pw"            "CA(R)"            
#> [202] "LX(L)"             "SLP(R)"            "AL-DM5(R)"        
#> [205] "SIP(L)"            "EBr2r4"            "NO(R)"            
#> [208] "FBl5"              "AL-DC3(R)"         "AL-VA1d(R)"       
#> [211] "PB(L3)"            "SAD"               "CRE(R)"           
#> [214] "EB"                "AL-DL4(L)"         "ICL(L)"           
#> [217] "a'2(R)"            "AL-DL5(L)"         "AL-VC1(R)"        
#> [220] "AL-VP3(R)"         "AB(L)"             "PRW"              
#> [223] "PB(R8)"            "AL-DC4(L)"         "AL-DP1m(R)"       
#> [226] "AL-VC5(R)"         "AL-VM2(R)"         "GA(R)"            
#> [229] "SPS(L)"            "LAL(L)"            "SNP(R)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "a'L(L)"            "b'1(R)"            "GA(R)"            
#>   [4] "FB"                "PB(L6)"            "CA(R)"            
#>   [7] "b'2(R)"            "PB(L7)"            "PVLP(R)"          
#>  [10] "PB(R5)"            "AVLP(R)"           "LX(L)"            
#>  [13] "bL(R)"             "VLNP(R)"           "WED(R)"           
#>  [16] "CRE(R)"            "ME(R)"             "EBr2r4"           
#>  [19] "vACA(R)"           "a1(R)"             "SAD"              
#>  [22] "SIP(L)"            "PB(R7)"            "LOP(R)"           
#>  [25] "PB(L2)"            "mALT(L)"           "SMP(L)"           
#>  [28] "aL(R)"             "gL(L)"             "RUB(L)"           
#>  [31] "SAD(-AMMC)"        "VES(L)"            "AB(L)"            
#>  [34] "INP"               "SPS(R)"            "PB(L8)"           
#>  [37] "a'L(R)"            "b1(R)"             "LX(R)"            
#>  [40] "PB(R3)"            "GOR(R)"            "PB(L5)"           
#>  [43] "NO2(R)"            "LAL(R)"            "a'3(R)"           
#>  [46] "EBr5"              "NO"                "CRE(L)"           
#>  [49] "FBl2"              "GF(R)"             "a'2(R)"           
#>  [52] "b'L(L)"            "OL(R)"             "SCL(L)"           
#>  [55] "PRW"               "PB(L9)"            "SNP(L)"           
#>  [58] "ATL(R)"            "b2(R)"             "ICL(R)"           
#>  [61] "CA(L)"             "g1(R)"             "BU(R)"            
#>  [64] "NO3(L)"            "LH(R)"             "SIP(R)"           
#>  [67] "CAN(R)"            "POC"               "EBr3am"           
#>  [70] "b'L(R)"            "EBr1"              "FLA(R)"           
#>  [73] "EPA(R)"            "FBl8"              "AME(R)"           
#>  [76] "PB(L4)"            "CX"                "aL(L)"            
#>  [79] "AMMC"              "PB(R6)"            "ROB(R)"           
#>  [82] "gL(R)"             "MB(+ACA)(R)"       "PB(R9)"           
#>  [85] "EBr3pw"            "FBl3"              "FBl6"             
#>  [88] "PB(L3)"            "MB(L)"             "EBr3d"            
#>  [91] "SNP(R)"            "AL(L)"             "IB"               
#>  [94] "ATL(L)"            "g3(R)"             "hemibrain"        
#>  [97] "LAL(-GA)(R)"       "EB"                "PB(R8)"           
#> [100] "LO(R)"             "FBl1"              "SLP(R)"           
#> [103] "AOT(R)"            "g4(R)"             "FBl5"             
#> [106] "GNG"               "MB(R)"             "dACA(R)"          
#> [109] "ICL(L)"            "FBl9"              "NO(L)"            
#> [112] "PED(R)"            "NO1(L)"            "g2(R)"            
#> [115] "BU(L)"             "mALT(R)"           "GC"               
#> [118] "PB(R1)"            "a3(R)"             "CRE(-ROB,-RUB)(R)"
#> [121] "SCL(R)"            "EBr6"              "VMNP"             
#> [124] "NO(R)"             "PB(R4)"            "AL(R)"            
#> [127] "NO2(L)"            "lACA(R)"           "PB(L1)"           
#> [130] "SPS(L)"            "FBl7"              "PB(R2)"           
#> [133] "SMP(R)"            "GOR(L)"            "PLP(R)"           
#> [136] "g5(R)"             "IPS(R)"            "VES(R)"           
#> [139] "AB(R)"             "RUB(R)"            "LAL(L)"           
#> [142] "AOTU(R)"           "NO3(R)"            "bL(L)"            
#> [145] "a'1(R)"            "EPA(L)"            "FBl4"             
#> [148] "PENP"              "a2(R)"             "NO1(R)"           
#> [151] "PB"               
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "VMNP"              "PB(L9)"            "FB-column3"       
#>   [4] "MB(R)"             "LAL(-GA)(R)"       "AL-VM7v(L)"       
#>   [7] "FBl1"              "SLP(R)"            "AL-VA1v(R)"       
#>  [10] "AL-VA6(R)"         "AL-VC3l(R)"        "AL-VM2(R)"        
#>  [13] "AL-VM5v(R)"        "CRE(-RUB)(L)"      "GOR(R)"           
#>  [16] "EBr1"              "FBl9"              "a3(R)"            
#>  [19] "AL-VL1(R)"         "SMP(R)"            "g3(R)"            
#>  [22] "PVLP(R)"           "FBl3"              "NO2(R)"           
#>  [25] "PB(L4)"            "AL-VC3m(R)"        "AL-VM4(R)"        
#>  [28] "AL-VP3(R)"         "AOT(R)"            "PB(R5)"           
#>  [31] "b2(R)"             "g2(R)"             "SCL(R)"           
#>  [34] "AL-DL4(L)"         "VES(R)"            "AL-DM4(R)"        
#>  [37] "SAD"               "BU(L)"             "AL-DA4m(L)"       
#>  [40] "AL-DM1(L)"         "NO(R)"             "AL-DM1(R)"        
#>  [43] "CA(L)"             "LAL(L)"            "SMP(L)"           
#>  [46] "INP"               "AL-DA3(R)"         "EPA(L)"           
#>  [49] "AL-DM6(R)"         "PB(L1)"            "AL-VA1d(R)"       
#>  [52] "AL-VA2(R)"         "AL-VC1(R)"         "GOR(L)"           
#>  [55] "PB(R8)"            "AL-DL3(R)"         "AL-VC4(R)"        
#>  [58] "a'1(R)"            "AL-DC1(R)"         "AL-DL4(R)"        
#>  [61] "CX"                "AL(R)"             "AL-DL1(R)"        
#>  [64] "CRE(-ROB,-RUB)(R)" "NO1(L)"            "AL-VM7d(L)"       
#>  [67] "a'3(R)"            "PENP"              "PED(R)"           
#>  [70] "EBr5"              "LO(R)"             "b'L(L)"           
#>  [73] "AL-VA7m(R)"        "AL-VM3(R)"         "PRW"              
#>  [76] "SIP(L)"            "PB(R2)"            "a'L(L)"           
#>  [79] "AL-DC1(L)"         "AL-DP1m(R)"        "b'2(R)"           
#>  [82] "hemibrain"         "CA(R)"             "AL-VA3(R)"        
#>  [85] "AL-VP1d(R)"        "b'L(R)"            "AL-VA6(L)"        
#>  [88] "AL-DC4(R)"         "AL-D(R)"           "AB(L)"            
#>  [91] "ME(R)"             "AL-DM3(L)"         "ROB(R)"           
#>  [94] "PB(L6)"            "PB(R9)"            "AL-V(R)"          
#>  [97] "AL-VM7v(R)"        "a2(R)"             "EBr6"             
#> [100] "FB"                "PB(L8)"            "AL-DA2(L)"        
#> [103] "bL(L)"             "FBl5"              "CAN(R)"           
#> [106] "NO1(R)"            "g4(R)"             "SNP(L)"           
#> [109] "bL(R)"             "vACA(R)"           "NO2(L)"           
#> [112] "dACA(R)"           "AL-VL2a(R)"        "AL-VP4(R)"        
#> [115] "RUB(L)"            "gL(L)"             "LH(R)"            
#> [118] "AL-DM5(R)"         "ICL(R)"            "AME(R)"           
#> [121] "AL-VA7l(R)"        "AL-VC2(R)"         "AL-DM5(L)"        
#> [124] "AL-DC3(R)"         "AL-DP1l(R)"        "aL(R)"            
#> [127] "FBl7"              "PB(R3)"            "VLNP(R)"          
#> [130] "SPS(R)"            "PB"                "b'1(R)"           
#> [133] "GA(R)"             "AL-DC4(L)"         "b1(R)"            
#> [136] "RUB(R)"            "PB(L7)"            "AL-DC2(R)"        
#> [139] "AL-VA5(R)"         "AMMC"              "GF(R)"            
#> [142] "POC"               "EBr3d"             "SNP(R)"           
#> [145] "EB"                "AL-DC2(L)"         "AL-D(L)"          
#> [148] "AL-DM2(L)"         "LX(L)"             "AL-VP5(R)"        
#> [151] "AL-VP2(R)"         "PB(R1)"            "CRE(L)"           
#> [154] "AL-DM2(R)"         "AB(R)"             "LX(R)"            
#> [157] "AL-DM6(L)"         "GNG"               "mALT(L)"          
#> [160] "SIP(R)"            "NO3(L)"            "MB(+ACA)(R)"      
#> [163] "AL-DM3(R)"         "g5(R)"             "IPS(R)"           
#> [166] "AL-VM1(R)"         "EBr2r4"            "NO"               
#> [169] "EBr3pw"            "FBl2"              "AL-DL2d(R)"       
#> [172] "SAD(-AMMC)"        "AL-VM5d(R)"        "SCL(L)"           
#> [175] "lACA(R)"           "a'L(R)"            "AL-DL2v(R)"       
#> [178] "ATL(R)"            "EBr3am"            "PB(L5)"           
#> [181] "PB(L3)"            "SPS(L)"            "EPA(R)"           
#> [184] "ICL(L)"            "FLA(R)"            "BU(R)"            
#> [187] "ATL(L)"            "GC"                "AL-DA3(L)"        
#> [190] "FBl4"              "AL-DA4l(R)"        "CRE(R)"           
#> [193] "AL-DM4(L)"         "gL(R)"             "FBl8"             
#> [196] "OL(R)"             "AL-VA4(R)"         "PB(R7)"           
#> [199] "FBl6"              "g1(R)"             "AL-DP1m(L)"       
#> [202] "PB(R4)"            "AL-DA4m(R)"        "AL-VP1m(R)"       
#> [205] "AOTU(R)"           "a'2(R)"            "AL(L)"            
#> [208] "NO(L)"             "VES(L)"            "AL-DA2(R)"        
#> [211] "AL-DL5(R)"         "AL-VC5(R)"         "AL-VM7d(R)"       
#> [214] "IB"                "PB(L2)"            "a1(R)"            
#> [217] "MB(L)"             "AVLP(R)"           "NO3(R)"           
#> [220] "PLP(R)"            "aL(L)"             "LOP(R)"           
#> [223] "PB(R6)"            "mALT(R)"           "WED(R)"           
#> [226] "AL-DL5(L)"         "LAL(R)"            "AL-DA1(R)"        
#> [229] "AL-VL2p(R)"        "AL-VP1l(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-DM5(L)"         "SNP(R)"            "PB(L7)"           
#>   [4] "SNP(L)"            "FBl9"              "hemibrain"        
#>   [7] "PB(R7)"            "PVLP(R)"           "FBl1"             
#>  [10] "LH(R)"             "AOT(R)"            "b'2(R)"           
#>  [13] "FLA(R)"            "PB(L9)"            "a3(R)"            
#>  [16] "AL-VA4(R)"         "AL-VA6(R)"         "EB"               
#>  [19] "AL-DA3(L)"         "AL-D(L)"           "AVLP(R)"          
#>  [22] "INP"               "AL(R)"             "EBr3pw"           
#>  [25] "LX(L)"             "AL-DL5(R)"         "g4(R)"            
#>  [28] "AL-DM6(L)"         "SIP(R)"            "NO(R)"            
#>  [31] "CRE(L)"            "GOR(R)"            "BU(R)"            
#>  [34] "CRE(-RUB)(L)"      "VES(R)"            "AL-VP4(R)"        
#>  [37] "LAL(L)"            "AL-DL3(R)"         "CRE(-ROB,-RUB)(R)"
#>  [40] "GA(R)"             "SPS(R)"            "AL-DA2(R)"        
#>  [43] "GF(R)"             "PB"                "ICL(L)"           
#>  [46] "AL-VP2(R)"         "POC"               "AB(L)"            
#>  [49] "WED(R)"            "aL(R)"             "FB"               
#>  [52] "a'L(R)"            "AL-DM1(R)"         "AL-VC1(R)"        
#>  [55] "vACA(R)"           "VMNP"              "gL(R)"            
#>  [58] "RUB(L)"            "AL-VP3(R)"         "a'3(R)"           
#>  [61] "PB(R4)"            "AL-DL2d(R)"        "AL-VA5(R)"        
#>  [64] "EBr3am"            "EBr3d"             "gL(L)"            
#>  [67] "MB(+ACA)(R)"       "FBl2"              "EBr5"             
#>  [70] "BU(L)"             "AL(L)"             "AME(R)"           
#>  [73] "CA(L)"             "AL-DC3(R)"         "PB(R5)"           
#>  [76] "AMMC"              "CX"                "AL-DA4m(R)"       
#>  [79] "AL-DL4(R)"         "AL-VM5v(R)"        "VLNP(R)"          
#>  [82] "AL-DC4(R)"         "AL-VL2p(R)"        "PB(R9)"           
#>  [85] "NO2(R)"            "OL(R)"             "AL-DM3(L)"        
#>  [88] "AL-V(R)"           "NO1(R)"            "PB(L4)"           
#>  [91] "mALT(L)"           "PB(L8)"            "AL-VM7d(L)"       
#>  [94] "ROB(R)"            "PB(R2)"            "a2(R)"            
#>  [97] "SAD(-AMMC)"        "AL-VM4(R)"         "AL-VC4(R)"        
#> [100] "PB(R6)"            "g3(R)"             "NO3(R)"           
#> [103] "b'L(R)"            "NO"                "AL-DM3(R)"        
#> [106] "AL-DC2(L)"         "FBl3"              "AL-DA1(R)"        
#> [109] "AL-VA1d(R)"        "a'1(R)"            "FBl5"             
#> [112] "AL-DM2(R)"         "AL-VA7m(R)"        "LOP(R)"           
#> [115] "ATL(R)"            "a1(R)"             "dACA(R)"          
#> [118] "a'2(R)"            "AL-DL2v(R)"        "AL-VM7v(R)"       
#> [121] "SCL(L)"            "PB(L3)"            "NO1(L)"           
#> [124] "SCL(R)"            "PLP(R)"            "AL-VP5(R)"        
#> [127] "SIP(L)"            "AL-DP1m(L)"        "LO(R)"            
#> [130] "AL-VP1d(R)"        "AL-DA3(R)"         "AL-DC1(R)"        
#> [133] "PRW"               "EBr6"              "g5(R)"            
#> [136] "AL-VM7d(R)"        "SMP(R)"            "EBr1"             
#> [139] "FBl6"              "EPA(L)"            "b'L(L)"           
#> [142] "SLP(R)"            "AL-DM1(L)"         "FB-column3"       
#> [145] "IPS(R)"            "SPS(L)"            "ICL(R)"           
#> [148] "CAN(R)"            "MB(R)"             "AL-DM5(R)"        
#> [151] "AL-VM1(R)"         "AL-DA4m(L)"        "AL-DL4(L)"        
#> [154] "AL-VA6(L)"         "NO3(L)"            "FBl7"             
#> [157] "g1(R)"             "AL-DA2(L)"         "AL-D(R)"          
#> [160] "AOTU(R)"           "PED(R)"            "ME(R)"            
#> [163] "AL-DM4(R)"         "NO2(L)"            "LAL(-GA)(R)"      
#> [166] "AL-DC4(L)"         "PB(R3)"            "AL-VC3m(R)"       
#> [169] "AL-VL1(R)"         "AL-VP1m(R)"        "EBr2r4"           
#> [172] "AL-DL5(L)"         "AL-DM2(L)"         "AL-VM7v(L)"       
#> [175] "LAL(R)"            "AL-DP1m(R)"        "mALT(R)"          
#> [178] "a'L(L)"            "AL-VP1l(R)"        "CA(R)"            
#> [181] "GNG"               "LX(R)"             "b'1(R)"           
#> [184] "lACA(R)"           "bL(L)"             "AL-VM2(R)"        
#> [187] "PB(L5)"            "SMP(L)"            "EPA(R)"           
#> [190] "AL-VA7l(R)"        "PB(L2)"            "AL-VL2a(R)"       
#> [193] "GC"                "FBl4"              "AB(R)"            
#> [196] "aL(L)"             "g2(R)"             "bL(R)"            
#> [199] "AL-VC3l(R)"        "PB(R8)"            "AL-DC2(R)"        
#> [202] "ATL(L)"            "PB(L1)"            "AL-DC1(L)"        
#> [205] "PENP"              "AL-VA2(R)"         "GOR(L)"           
#> [208] "AL-DM4(L)"         "PB(R1)"            "AL-DA4l(R)"       
#> [211] "AL-DP1l(R)"        "SAD"               "b2(R)"            
#> [214] "VES(L)"            "AL-DL1(R)"         "RUB(R)"           
#> [217] "PB(L6)"            "NO(L)"             "AL-VC2(R)"        
#> [220] "AL-VC5(R)"         "IB"                "CRE(R)"           
#> [223] "MB(L)"             "FBl8"              "b1(R)"            
#> [226] "AL-DM6(R)"         "AL-VA1v(R)"        "AL-VA3(R)"        
#> [229] "AL-VM3(R)"         "AL-VM5d(R)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                  
#> [1,] "POST" "/api"                                
#> [2,] "GET"  "/api/npexplorer/findneurons"         
#> [3,] "POST" "/api/v:ver/npexplorer/completeness"  
#> [4,] "GET"  "/api/raw/keyvalue/key/:instance/:key"
#> [5,] "GET"  "/api/available"                      
#> [6,] "GET"  "/oauth2callback"                     
#>      name                                                                                       
#> [1,] "github.com/labstack/echo.glob..func1"                                                     
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getFindNeurons-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/raw/keyvalue.masterAPI.getKV-fm"          
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.SetupRoutes.func1"                        
#> [6,] "github.com/janelia-flyem/echo-secure.oauthCallbackHandler"                                

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }