Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(L9)"            "AVLP(R)"           "b1(R)"            
#>   [4] "NO2(R)"            "NO2(L)"            "AOTU(R)"          
#>   [7] "b2(R)"             "NO1(R)"            "g2(R)"            
#>  [10] "PB(R6)"            "MB(L)"             "LX(R)"            
#>  [13] "BU(R)"             "CRE(L)"            "LAL(-GA)(R)"      
#>  [16] "mALT(R)"           "SPS(L)"            "NO(L)"            
#>  [19] "GA(R)"             "hemibrain"         "PB(R8)"           
#>  [22] "a'2(R)"            "LO(R)"             "a1(R)"            
#>  [25] "b'L(L)"            "PB"                "SNP(L)"           
#>  [28] "RUB(L)"            "NO3(R)"            "FB"               
#>  [31] "EBr1"              "dACA(R)"           "gL(L)"            
#>  [34] "GF(R)"             "lACA(R)"           "PED(R)"           
#>  [37] "ATL(R)"            "EPA(R)"            "AL(L)"            
#>  [40] "LH(R)"             "PB(L3)"            "CAN(R)"           
#>  [43] "a'L(L)"            "ATL(L)"            "EBr3d"            
#>  [46] "g4(R)"             "b'1(R)"            "GOR(L)"           
#>  [49] "WED(R)"            "POC"               "a'L(R)"           
#>  [52] "PB(R9)"            "SAD(-AMMC)"        "FBl7"             
#>  [55] "IPS(R)"            "aL(R)"             "BU(L)"            
#>  [58] "EB"                "CX"                "LX(L)"            
#>  [61] "bL(R)"             "MB(R)"             "INP"              
#>  [64] "PB(L1)"            "PB(R3)"            "AB(R)"            
#>  [67] "vACA(R)"           "EBr6"              "NO(R)"            
#>  [70] "PB(R2)"            "PLP(R)"            "AOT(R)"           
#>  [73] "PRW"               "SNP(R)"            "LAL(R)"           
#>  [76] "EBr3pw"            "FBl4"              "SLP(R)"           
#>  [79] "RUB(R)"            "PB(L5)"            "g3(R)"            
#>  [82] "a2(R)"             "SCL(R)"            "g5(R)"            
#>  [85] "FBl2"              "PB(L4)"            "AMMC"             
#>  [88] "NO1(L)"            "b'L(R)"            "FLA(R)"           
#>  [91] "PB(L7)"            "SCL(L)"            "CRE(-ROB,-RUB)(R)"
#>  [94] "PB(R7)"            "EBr2r4"            "NO"               
#>  [97] "PB(L8)"            "a'3(R)"            "LAL(L)"           
#> [100] "SPS(R)"            "CRE(R)"            "PB(R1)"           
#> [103] "OL(R)"             "AL(R)"             "EBr3am"           
#> [106] "a'1(R)"            "FBl1"              "SMP(R)"           
#> [109] "ROB(R)"            "PB(L6)"            "NO3(L)"           
#> [112] "VES(R)"            "FBl3"              "b'2(R)"           
#> [115] "SAD"               "SIP(R)"            "ICL(R)"           
#> [118] "PB(R4)"            "GNG"               "VES(L)"           
#> [121] "SIP(L)"            "gL(R)"             "PVLP(R)"          
#> [124] "ME(R)"             "FBl5"              "LOP(R)"           
#> [127] "FBl6"              "FBl9"              "CA(R)"            
#> [130] "CA(L)"             "g1(R)"             "EBr5"             
#> [133] "MB(+ACA)(R)"       "PB(R5)"            "IB"               
#> [136] "mALT(L)"           "AME(R)"            "aL(L)"            
#> [139] "EPA(L)"            "a3(R)"             "ICL(L)"           
#> [142] "FBl8"              "GOR(R)"            "PB(L2)"           
#> [145] "PENP"              "SMP(L)"            "AB(L)"            
#> [148] "VLNP(R)"           "GC"                "VMNP"             
#> [151] "bL(L)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-VA6(R)"         "AL-VA7l(R)"        "AL-VL2a(R)"       
#>   [4] "LOP(R)"            "PB(L6)"            "IPS(R)"           
#>   [7] "a'L(R)"            "AL-DC4(L)"         "PB(R3)"           
#>  [10] "AL-DA2(R)"         "FB"                "bL(L)"            
#>  [13] "FBl9"              "AL(R)"             "AL-DC1(R)"        
#>  [16] "VMNP"              "CRE(L)"            "AMMC"             
#>  [19] "AL-DM3(L)"         "VES(R)"            "PVLP(R)"          
#>  [22] "a'3(R)"            "EBr6"              "SPS(L)"           
#>  [25] "aL(L)"             "AL-DM5(R)"         "AL-VP1m(R)"       
#>  [28] "SNP(R)"            "ICL(R)"            "AL-VA2(R)"        
#>  [31] "WED(R)"            "PB(L8)"            "CA(R)"            
#>  [34] "AL-VA4(R)"         "AL-DL4(R)"         "AL-VP1d(R)"       
#>  [37] "AL-DA1(R)"         "AL-DM3(R)"         "AL-VA7m(R)"       
#>  [40] "SIP(L)"            "AL-VA1d(R)"        "PB(L4)"           
#>  [43] "AL-VA5(R)"         "a'2(R)"            "AL-DM4(L)"        
#>  [46] "NO3(L)"            "FBl3"              "AB(R)"            
#>  [49] "PB(R4)"            "AL-VC2(R)"         "PB(R5)"           
#>  [52] "PENP"              "EBr2r4"            "NO(R)"            
#>  [55] "INP"               "AL-VM7d(R)"        "AL-DA3(R)"        
#>  [58] "FBl1"              "AL-DM1(R)"         "SAD"              
#>  [61] "GOR(L)"            "g1(R)"             "NO1(L)"           
#>  [64] "AL-DA2(L)"         "FB-column3"        "AOT(R)"           
#>  [67] "SNP(L)"            "GOR(R)"            "AL-VM5v(R)"       
#>  [70] "AL-V(R)"           "EBr3pw"            "FBl6"             
#>  [73] "ATL(R)"            "LAL(R)"            "AL-VC5(R)"        
#>  [76] "AL-VM4(R)"         "EB"                "PB"               
#>  [79] "LX(R)"             "FBl5"              "AL-DM2(R)"        
#>  [82] "RUB(R)"            "lACA(R)"           "GC"               
#>  [85] "AL-DP1m(R)"        "AL-VC3m(R)"        "AL-VP3(R)"        
#>  [88] "b'2(R)"            "g2(R)"             "AL-DP1m(L)"       
#>  [91] "NO(L)"             "LAL(L)"            "b2(R)"            
#>  [94] "AL-VP1l(R)"        "PB(L5)"            "AL-DP1l(R)"       
#>  [97] "AL-VA3(R)"         "AL-VM7v(R)"        "mALT(L)"          
#> [100] "AL-DL4(L)"         "AL-VM7v(L)"        "g3(R)"            
#> [103] "CAN(R)"            "CRE(-ROB,-RUB)(R)" "SMP(L)"           
#> [106] "g4(R)"             "AL-DM4(R)"         "AL-VM3(R)"        
#> [109] "AL-DC1(L)"         "NO2(R)"            "AL-DM6(R)"        
#> [112] "AL-DC4(R)"         "POC"               "EBr3d"            
#> [115] "BU(R)"             "AL(L)"             "AL-DM6(L)"        
#> [118] "AL-VM7d(L)"        "bL(R)"             "NO2(L)"           
#> [121] "NO3(R)"            "GF(R)"             "GA(R)"            
#> [124] "CRE(R)"            "ME(R)"             "AL-VA1v(R)"       
#> [127] "EBr3am"            "LAL(-GA)(R)"       "aL(R)"            
#> [130] "NO"                "SAD(-AMMC)"        "PED(R)"           
#> [133] "AL-DC2(R)"         "IB"                "PRW"              
#> [136] "LX(L)"             "AL-DA4l(R)"        "AB(L)"            
#> [139] "SMP(R)"            "SIP(R)"            "SCL(L)"           
#> [142] "PB(R6)"            "EBr5"              "FBl2"             
#> [145] "AL-VC1(R)"         "AL-VM1(R)"         "NO1(R)"           
#> [148] "AL-DA3(L)"         "AL-DM1(L)"         "AL-D(L)"          
#> [151] "AL-VC4(R)"         "a2(R)"             "dACA(R)"          
#> [154] "g5(R)"             "OL(R)"             "AL-DA4m(R)"       
#> [157] "AL-VP2(R)"         "vACA(R)"           "MB(L)"            
#> [160] "AL-DL1(R)"         "AL-DL5(R)"         "AL-DM5(L)"        
#> [163] "RUB(L)"            "AL-VM5d(R)"        "BU(L)"            
#> [166] "EPA(R)"            "b'1(R)"            "AL-VM2(R)"        
#> [169] "PB(L3)"            "PB(R7)"            "mALT(R)"          
#> [172] "b'L(R)"            "PB(L1)"            "AL-VC3l(R)"       
#> [175] "SPS(R)"            "a'L(L)"            "AL-DC3(R)"        
#> [178] "CA(L)"             "hemibrain"         "AL-DC2(L)"        
#> [181] "ROB(R)"            "LO(R)"             "AL-VL2p(R)"       
#> [184] "AOTU(R)"           "MB(R)"             "PB(R8)"           
#> [187] "gL(R)"             "AVLP(R)"           "a3(R)"            
#> [190] "AL-DL3(R)"         "a1(R)"             "AL-VP5(R)"        
#> [193] "EBr1"              "b'L(L)"            "AL-DL2v(R)"       
#> [196] "AL-D(R)"           "SCL(R)"            "PLP(R)"           
#> [199] "PB(L9)"            "AL-DL2d(R)"        "PB(L2)"           
#> [202] "FLA(R)"            "AL-DA4m(L)"        "AL-VA6(L)"        
#> [205] "a'1(R)"            "FBl4"              "SLP(R)"           
#> [208] "AL-DM2(L)"         "PB(R9)"            "ICL(L)"           
#> [211] "AL-VL1(R)"         "FBl7"              "CX"               
#> [214] "AL-DL5(L)"         "CRE(-RUB)(L)"      "LH(R)"            
#> [217] "b1(R)"             "VLNP(R)"           "EPA(L)"           
#> [220] "ATL(L)"            "gL(L)"             "PB(R1)"           
#> [223] "FBl8"              "VES(L)"            "PB(R2)"           
#> [226] "GNG"               "AL-VP4(R)"         "PB(L7)"           
#> [229] "MB(+ACA)(R)"       "AME(R)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                        
#> [1,] "POST" "/api/raw/cypher/transaction/:id/commit"    
#> [2,] "GET"  "/token"                                    
#> [3,] "GET"  "/api/v:ver/npexplorer/neuronmeta"          
#> [4,] "POST" "/api/npexplorer/roisinneuron"              
#> [5,] "GET"  "/oauth2callback"                           
#> [6,] "GET"  "/api/v:ver/skeletons/skeleton/:dataset/:id"
#>      name                                                                                        
#> [1,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [2,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"
#> [5,] "github.com/janelia-flyem/echo-secure.oauthCallbackHandler"                                 
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }