Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "bL(R)"             "GOR(R)"            "SPS(R)"           
#>   [4] "FB-column3"        "bL(L)"             "AMMC"             
#>   [7] "AL-DL1(R)"         "AL-VM2(R)"         "AOT(R)"           
#>  [10] "aL(R)"             "EBr5"              "PB(R1)"           
#>  [13] "b'L(L)"            "AL-VM3(R)"         "BU(L)"            
#>  [16] "SPS(L)"            "LX(R)"             "b1(R)"            
#>  [19] "VLNP(R)"           "EBr3am"            "SIP(L)"           
#>  [22] "PB"                "INP"               "AL-VL1(R)"        
#>  [25] "dACA(R)"           "VMNP"              "AL-VP2(R)"        
#>  [28] "AB(L)"             "b'1(R)"            "a1(R)"            
#>  [31] "AOTU(R)"           "FB"                "VES(L)"           
#>  [34] "g1(R)"             "AL-D(L)"           "FBl1"             
#>  [37] "FBl5"              "AL-DA4m(R)"        "AL-DM4(R)"        
#>  [40] "AL-VC1(R)"         "BU(R)"             "MB(+ACA)(R)"      
#>  [43] "AL-VA1d(R)"        "AL-DC2(L)"         "AL-DP1l(R)"       
#>  [46] "AL-VC2(R)"         "PB(R6)"            "SAD"              
#>  [49] "AL-VP3(R)"         "AL-DM3(R)"         "AL-VA2(R)"        
#>  [52] "AL-VA6(R)"         "SMP(R)"            "PB(R2)"           
#>  [55] "PB(L4)"            "AL-DM1(R)"         "AL-V(R)"          
#>  [58] "AL-DM2(L)"         "FBl6"              "AL-VM7v(L)"       
#>  [61] "ICL(R)"            "SLP(R)"            "AL(R)"            
#>  [64] "IB"                "EB"                "AL-DM3(L)"        
#>  [67] "AL-DC1(L)"         "AL-VL2p(R)"        "AL-VM5v(R)"       
#>  [70] "PB(R7)"            "b2(R)"             "EBr1"             
#>  [73] "LAL(R)"            "FBl4"              "AL-VC3m(R)"       
#>  [76] "CA(L)"             "PED(R)"            "PB(L9)"           
#>  [79] "PVLP(R)"           "AL-VC5(R)"         "EBr3pw"           
#>  [82] "AL-VM5d(R)"        "NO2(L)"            "AL-DA3(L)"        
#>  [85] "gL(R)"             "FBl9"              "CX"               
#>  [88] "PLP(R)"            "AL-DM1(L)"         "ATL(R)"           
#>  [91] "AL-DA2(R)"         "LOP(R)"            "AL-DM6(L)"        
#>  [94] "IB112"             "NO(R)"             "MB(R)"            
#>  [97] "AL-DC4(R)"         "AL-VM7d(R)"        "PRW"              
#> [100] "ME(R)"             "AL-VA4(R)"         "AL-VC3l(R)"       
#> [103] "AL-VP1d(R)"        "FBl8"              "AL-DL2v(R)"       
#> [106] "AVLP(R)"           "AL-DP1m(L)"        "FBl3"             
#> [109] "a'1(R)"            "AL-DA4l(R)"        "SCL(R)"           
#> [112] "EBr3d"             "IPS(R)"            "EPA(L)"           
#> [115] "AL-VM4(R)"         "WED(R)"            "NO3(R)"           
#> [118] "POC"               "SNP(L)"            "AL-DA2(L)"        
#> [121] "AL-DM4(L)"         "AL-VA5(R)"         "AL-VM7v(R)"       
#> [124] "PB(L2)"            "PB(L1)"            "a'L(R)"           
#> [127] "PB(R3)"            "aL(L)"             "GNG"              
#> [130] "OL(R)"             "AL-DP1m(R)"        "a'3(R)"           
#> [133] "GA(R)"             "VES(R)"            "mALT(L)"          
#> [136] "NO(L)"             "vACA(R)"           "ICL(L)"           
#> [139] "FBl7"              "b'2(R)"            "AME(R)"           
#> [142] "AL-DL5(R)"         "AL-VC4(R)"         "EBr6"             
#> [145] "AL-DL2d(R)"        "LH(R)"             "AL-DL4(R)"        
#> [148] "AL-VA7m(R)"        "AL-VP4(R)"         "LAL(-GA)(R)"      
#> [151] "AL(L)"             "SNP(R)"            "AL-D(R)"          
#> [154] "PB(L8)"            "SIP(R)"            "PB(L6)"           
#> [157] "SCL(L)"            "b'L(R)"            "ROB(R)"           
#> [160] "FBl2"              "CRE(-ROB,-RUB)(R)" "GF(R)"            
#> [163] "CRE(R)"            "AL-DA4m(L)"        "PB(L5)"           
#> [166] "PB(R5)"            "a'L(L)"            "AL-VA1v(R)"       
#> [169] "a3(R)"             "lACA(R)"           "NO2(R)"           
#> [172] "AL-DL5(L)"         "CRE(L)"            "PB(L7)"           
#> [175] "RUB(L)"            "ATL(L)"            "a2(R)"            
#> [178] "MB(L)"             "EBr2r4"            "g5(R)"            
#> [181] "AL-DC2(R)"         "LAL(L)"            "SMP(L)"           
#> [184] "LX(L)"             "AL-DA1(R)"         "GOR(L)"           
#> [187] "mALT(R)"           "PB(R9)"            "g3(R)"            
#> [190] "a'2(R)"            "AL-DM2(R)"         "AL-VP1l(R)"       
#> [193] "AL-DC4(L)"         "AL-DC3(R)"         "AL-VM1(R)"        
#> [196] "SAD(-AMMC)"        "AL-DL3(R)"         "RUB(R)"           
#> [199] "PENP"              "AL-DL4(L)"         "g4(R)"            
#> [202] "PB(L3)"            "EPA(R)"            "CA(R)"            
#> [205] "PB(R8)"            "LO(R)"             "AL-VA6(L)"        
#> [208] "NO"                "CRE(-RUB)(L)"      "AB(R)"            
#> [211] "PB(R4)"            "AL-DM6(R)"         "NO1(L)"           
#> [214] "gL(L)"             "CAN(R)"            "AL-DM5(R)"        
#> [217] "AL-VA7l(R)"        "hemibrain"         "FLA(R)"           
#> [220] "NO1(R)"            "NO3(L)"            "AL-DA3(R)"        
#> [223] "AL-VA3(R)"         "AL-VL2a(R)"        "AL-VP5(R)"        
#> [226] "GC"                "g2(R)"             "AL-DC1(R)"        
#> [229] "AL-VP1m(R)"        "AL-DM5(L)"         "AL-VM7d(L)"       
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PENP"              "LAL(-GA)(R)"       "NO3(R)"           
#>   [4] "EBr5"              "SAD(-AMMC)"        "FBl4"             
#>   [7] "LAL(L)"            "SMP(L)"            "SPS(R)"           
#>  [10] "PB(R9)"            "GOR(R)"            "EBr3am"           
#>  [13] "b'L(R)"            "PB(L1)"            "PB(L4)"           
#>  [16] "PB(L2)"            "a'3(R)"            "BU(R)"            
#>  [19] "AOTU(R)"           "GA(R)"             "SIP(L)"           
#>  [22] "a1(R)"             "b2(R)"             "NO1(R)"           
#>  [25] "PB(L9)"            "FBl3"              "AME(R)"           
#>  [28] "AOT(R)"            "CA(L)"             "GF(R)"            
#>  [31] "g3(R)"             "PB(L3)"            "GC"               
#>  [34] "FBl7"              "hemibrain"         "LX(R)"            
#>  [37] "ROB(R)"            "AVLP(R)"           "ICL(R)"           
#>  [40] "OL(R)"             "PRW"               "MB(+ACA)(R)"      
#>  [43] "mALT(R)"           "GNG"               "IB"               
#>  [46] "PB(L7)"            "a3(R)"             "LAL(R)"           
#>  [49] "LO(R)"             "IPS(R)"            "AL(R)"            
#>  [52] "PB(R7)"            "NO3(L)"            "FBl5"             
#>  [55] "RUB(R)"            "SMP(R)"            "SCL(L)"           
#>  [58] "mALT(L)"           "CRE(R)"            "BU(L)"            
#>  [61] "dACA(R)"           "a'1(R)"            "a'2(R)"           
#>  [64] "gL(L)"             "LH(R)"             "AB(L)"            
#>  [67] "PED(R)"            "EBr6"              "PB(L8)"           
#>  [70] "VMNP"              "EBr3d"             "CRE(L)"           
#>  [73] "AB(R)"             "NO2(L)"            "FLA(R)"           
#>  [76] "PB(R2)"            "NO(L)"             "SLP(R)"           
#>  [79] "PB(R8)"            "NO(R)"             "PB(R3)"           
#>  [82] "MB(R)"             "POC"               "CX"               
#>  [85] "AL(L)"             "FBl2"              "FBl8"             
#>  [88] "PB"                "ME(R)"             "PB(L6)"           
#>  [91] "SNP(L)"            "PVLP(R)"           "a'L(L)"           
#>  [94] "vACA(R)"           "ATL(L)"            "aL(R)"            
#>  [97] "g2(R)"             "SCL(R)"            "FBl6"             
#> [100] "EB"                "ICL(L)"            "b'2(R)"           
#> [103] "EBr2r4"            "PB(L5)"            "NO1(L)"           
#> [106] "SNP(R)"            "EBr3pw"            "CA(R)"            
#> [109] "ATL(R)"            "b1(R)"             "RUB(L)"           
#> [112] "b'L(L)"            "AMMC"              "MB(L)"            
#> [115] "EPA(R)"            "bL(L)"             "g5(R)"            
#> [118] "lACA(R)"           "SPS(L)"            "gL(R)"            
#> [121] "FBl9"              "LX(L)"             "INP"              
#> [124] "PB(R4)"            "VLNP(R)"           "WED(R)"           
#> [127] "EBr1"              "CAN(R)"            "aL(L)"            
#> [130] "VES(L)"            "GOR(L)"            "PB(R1)"           
#> [133] "FBl1"              "bL(R)"             "PB(R5)"           
#> [136] "PB(R6)"            "a'L(R)"            "g4(R)"            
#> [139] "VES(R)"            "b'1(R)"            "CRE(-ROB,-RUB)(R)"
#> [142] "a2(R)"             "SAD"               "g1(R)"            
#> [145] "FB"                "PLP(R)"            "SIP(R)"           
#> [148] "NO"                "NO2(R)"            "EPA(L)"           
#> [151] "LOP(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "GOR(R)"            "AL-VM4(R)"         "VLNP(R)"          
#>   [4] "EBr3pw"            "AL-DA2(R)"         "PB(R1)"           
#>   [7] "LO(R)"             "AL-DP1l(R)"        "PB(R5)"           
#>  [10] "SMP(L)"            "gL(R)"             "SIP(L)"           
#>  [13] "EBr3am"            "CRE(-ROB,-RUB)(R)" "PB(L1)"           
#>  [16] "EBr2r4"            "AL-DL5(R)"         "PB(L2)"           
#>  [19] "ICL(R)"            "b1(R)"             "CAN(R)"           
#>  [22] "a'L(L)"            "FBl7"              "a3(R)"            
#>  [25] "AL-VA1v(R)"        "AB(L)"             "PB(R3)"           
#>  [28] "EBr1"              "BU(L)"             "PB(L6)"           
#>  [31] "FBl3"              "RUB(R)"            "BU(R)"            
#>  [34] "g5(R)"             "INP"               "AL-VC3l(R)"       
#>  [37] "SCL(R)"            "FBl9"              "AL-DA4m(R)"       
#>  [40] "PB(R9)"            "bL(L)"             "AL-DL2v(R)"       
#>  [43] "vACA(R)"           "PB(R6)"            "EB"               
#>  [46] "AL-DM3(L)"         "AL-DL1(R)"         "AL-VA5(R)"        
#>  [49] "FLA(R)"            "gL(L)"             "a'1(R)"           
#>  [52] "AL-DM4(R)"         "MB(+ACA)(R)"       "AME(R)"           
#>  [55] "AL-VL2a(R)"        "AL-VM5d(R)"        "a2(R)"            
#>  [58] "PB(R7)"            "AL-DC2(L)"         "AL-DL4(L)"        
#>  [61] "CRE(L)"            "AL-VA6(R)"         "AL-VM5v(R)"       
#>  [64] "AL-DM6(L)"         "CRE(-RUB)(L)"      "IPS(R)"           
#>  [67] "PB(L3)"            "hemibrain"         "LAL(-GA)(R)"      
#>  [70] "WED(R)"            "PVLP(R)"           "AL-VA3(R)"        
#>  [73] "AL-VM7d(R)"        "FBl6"              "AL-DC3(R)"        
#>  [76] "AL-DM5(R)"         "AL-VM3(R)"         "AL-VP2(R)"        
#>  [79] "AOT(R)"            "SPS(R)"            "SNP(L)"           
#>  [82] "bL(R)"             "AL-VA1d(R)"        "GA(R)"            
#>  [85] "b'L(L)"            "PED(R)"            "AL-DM4(L)"        
#>  [88] "GNG"               "AL-DM6(R)"         "FBl1"             
#>  [91] "AL-VM2(R)"         "PB(L4)"            "AMMC"             
#>  [94] "b'2(R)"            "b'L(R)"            "AL-DC1(L)"        
#>  [97] "AL-DL5(L)"         "RUB(L)"            "SLP(R)"           
#> [100] "AL-DM3(R)"         "PB(L5)"            "AL-DC4(R)"        
#> [103] "NO1(L)"            "g4(R)"             "PB(R2)"           
#> [106] "MB(R)"             "AL-DM2(R)"         "CX"               
#> [109] "SPS(L)"            "ME(R)"             "EPA(R)"           
#> [112] "FBl4"              "NO"                "NO3(L)"           
#> [115] "FB"                "EBr3d"             "SIP(R)"           
#> [118] "dACA(R)"           "NO2(R)"            "AOTU(R)"          
#> [121] "MB(L)"             "SAD(-AMMC)"        "PB(L8)"           
#> [124] "AL-DA4m(L)"        "IB"                "EBr6"             
#> [127] "NO(R)"             "AL-VA7l(R)"        "AL-VA7m(R)"       
#> [130] "AL-VC3m(R)"        "b'1(R)"            "AL-DC4(L)"        
#> [133] "AL-DM2(L)"         "VES(L)"            "AL-DL2d(R)"       
#> [136] "AVLP(R)"           "AL-DA3(R)"         "AL-VP4(R)"        
#> [139] "SCL(L)"            "AL-D(R)"           "AL-VC2(R)"        
#> [142] "AL-VM7v(R)"        "GF(R)"             "GOR(L)"           
#> [145] "AL-D(L)"           "aL(L)"             "POC"              
#> [148] "g3(R)"             "aL(R)"             "AL-DL4(R)"        
#> [151] "AL-VP1m(R)"        "CA(R)"             "AL-DP1m(R)"       
#> [154] "NO2(L)"            "PB(L7)"            "AL-DC1(R)"        
#> [157] "AL-VP1d(R)"        "LOP(R)"            "PENP"             
#> [160] "g1(R)"             "EBr5"              "PB(R8)"           
#> [163] "AL-DM5(L)"         "EPA(L)"            "PRW"              
#> [166] "LX(R)"             "PB(L9)"            "LAL(R)"           
#> [169] "NO(L)"             "AL-DC2(R)"         "AL-VC1(R)"        
#> [172] "AL-V(R)"           "FBl2"              "AL-VA4(R)"        
#> [175] "SAD"               "CRE(R)"            "a'L(R)"           
#> [178] "AL-DP1m(L)"        "ATL(R)"            "AL-VC5(R)"        
#> [181] "AL-VL2p(R)"        "SMP(R)"            "FBl8"             
#> [184] "AL-VC4(R)"         "AL-VM1(R)"         "a1(R)"            
#> [187] "VMNP"              "AL(R)"             "AL-VA2(R)"        
#> [190] "LX(L)"             "AL-VP1l(R)"        "GC"               
#> [193] "AB(R)"             "AL-DM1(R)"         "AL-VM7v(L)"       
#> [196] "LH(R)"             "AL-DL3(R)"         "AL-DA2(L)"        
#> [199] "AL-DM1(L)"         "PB"                "a'2(R)"           
#> [202] "SNP(R)"            "PB(R4)"            "PLP(R)"           
#> [205] "ROB(R)"            "AL-VP5(R)"         "mALT(L)"          
#> [208] "b2(R)"             "AL(L)"             "AL-DA1(R)"        
#> [211] "a'3(R)"            "AL-DA3(L)"         "AL-VA6(L)"        
#> [214] "ICL(L)"            "AL-VP3(R)"         "lACA(R)"          
#> [217] "NO1(R)"            "ATL(L)"            "AL-DA4l(R)"       
#> [220] "AL-VL1(R)"         "CA(L)"             "LAL(L)"           
#> [223] "mALT(R)"           "g2(R)"             "OL(R)"            
#> [226] "FB-column3"        "FBl5"              "NO3(R)"           
#> [229] "AL-VM7d(L)"        "VES(R)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "NO(R)"             "VES(R)"            "AL-DC1(R)"        
#>   [4] "AL-DM3(R)"         "AL-VA2(R)"         "NO1(R)"           
#>   [7] "AL-DP1m(L)"        "GA(R)"             "AL-DL5(L)"        
#>  [10] "NO3(L)"            "aL(L)"             "PB(L4)"           
#>  [13] "AL-DM5(R)"         "PB(R7)"            "a'L(R)"           
#>  [16] "SIP(R)"            "PB(L7)"            "GF(R)"            
#>  [19] "PB"                "PVLP(R)"           "PB(R2)"           
#>  [22] "AL-DA3(R)"         "AL-VM2(R)"         "NO2(L)"           
#>  [25] "g3(R)"             "PB(L8)"            "bL(L)"            
#>  [28] "NO(L)"             "a'L(L)"            "AL-DM1(R)"        
#>  [31] "AOTU(R)"           "ATL(R)"            "AMMC"             
#>  [34] "GOR(L)"            "AL-VM4(R)"         "PENP"             
#>  [37] "LAL(L)"            "PED(R)"            "AL-DL5(R)"        
#>  [40] "AL-VA1d(R)"        "AL-VA5(R)"         "AL-DC1(L)"        
#>  [43] "b1(R)"             "AL-VC4(R)"         "PB(L5)"           
#>  [46] "CA(L)"             "BU(R)"             "AL-DM3(L)"        
#>  [49] "FBl1"              "bL(R)"             "AL-DC2(R)"        
#>  [52] "AL-DL2v(R)"        "AL-VP5(R)"         "AL-DC2(L)"        
#>  [55] "ICL(R)"            "AL-DP1l(R)"        "a1(R)"            
#>  [58] "NO1(L)"            "SPS(L)"            "AL-DA2(R)"        
#>  [61] "AB(L)"             "PB(R6)"            "AL-DA3(L)"        
#>  [64] "GNG"               "VES(L)"            "AL-DC4(R)"        
#>  [67] "NO"                "LO(R)"             "IPS(R)"           
#>  [70] "OL(R)"             "EBr3am"            "AL-VM7v(L)"       
#>  [73] "PB(R3)"            "AOT(R)"            "ATL(L)"           
#>  [76] "AL-DM5(L)"         "FBl3"              "AL-VL2p(R)"       
#>  [79] "AL-DA4m(L)"        "a3(R)"             "AL-DL4(R)"        
#>  [82] "AL-D(R)"           "AL-VP4(R)"         "VLNP(R)"          
#>  [85] "MB(L)"             "SCL(R)"            "SLP(R)"           
#>  [88] "RUB(R)"            "AL-VP2(R)"         "SIP(L)"           
#>  [91] "AL-VC1(R)"         "b'2(R)"            "FBl7"             
#>  [94] "LAL(R)"            "AL-VC3l(R)"        "AL-VM5d(R)"       
#>  [97] "a'2(R)"            "FBl5"              "LH(R)"            
#> [100] "vACA(R)"           "FB-column3"        "b'L(L)"           
#> [103] "AL-DP1m(R)"        "LOP(R)"            "aL(R)"            
#> [106] "EBr6"              "LX(L)"             "EBr1"             
#> [109] "AL(L)"             "EBr3pw"            "AL-DM2(L)"        
#> [112] "AL-DM4(L)"         "AL-VA6(L)"         "AL-DA4l(R)"       
#> [115] "AL-VM1(R)"         "AL-VP1d(R)"        "a2(R)"            
#> [118] "AL-DL1(R)"         "WED(R)"            "EBr5"             
#> [121] "ME(R)"             "AL-DM1(L)"         "AL-DA4m(R)"       
#> [124] "FB"                "gL(L)"             "EB"               
#> [127] "AL-VM7d(L)"        "INP"               "FBl2"             
#> [130] "AL-VP1l(R)"        "g4(R)"             "CAN(R)"           
#> [133] "CX"                "AL-V(R)"           "SAD(-AMMC)"       
#> [136] "mALT(L)"           "hemibrain"         "PRW"              
#> [139] "SPS(R)"            "VMNP"              "AL-DM6(L)"        
#> [142] "a'1(R)"            "CRE(-RUB)(L)"      "AB(R)"            
#> [145] "AL-DL2d(R)"        "AL-VM5v(R)"        "AL-DL4(L)"        
#> [148] "PB(L3)"            "PB(R8)"            "SCL(L)"           
#> [151] "AVLP(R)"           "FBl4"              "b2(R)"            
#> [154] "AL-DA2(L)"         "ICL(L)"            "b'1(R)"           
#> [157] "AL-DM4(R)"         "g1(R)"             "PB(R9)"           
#> [160] "AL-VP1m(R)"        "PB(R5)"            "g2(R)"            
#> [163] "ROB(R)"            "AL-VA7m(R)"        "AL-VL2a(R)"       
#> [166] "PB(L1)"            "SNP(R)"            "SNP(L)"           
#> [169] "AL-VA1v(R)"        "CRE(R)"            "FLA(R)"           
#> [172] "EBr3d"             "PB(R1)"            "EBr2r4"           
#> [175] "FBl8"              "RUB(L)"            "AL-VC3m(R)"       
#> [178] "SMP(R)"            "GC"                "CA(R)"            
#> [181] "MB(R)"             "AL-VC2(R)"         "IB"               
#> [184] "AME(R)"            "NO2(R)"            "AL-DM2(R)"        
#> [187] "AL-VL1(R)"         "PB(L2)"            "lACA(R)"          
#> [190] "AL-D(L)"           "EPA(L)"            "PB(R4)"           
#> [193] "AL-VA7l(R)"        "MB(+ACA)(R)"       "FBl6"             
#> [196] "NO3(R)"            "BU(L)"             "AL-DC4(L)"        
#> [199] "PB(L6)"            "AL-DL3(R)"         "SAD"              
#> [202] "dACA(R)"           "a'3(R)"            "b'L(R)"           
#> [205] "gL(R)"             "AL-DA1(R)"         "AL-VC5(R)"        
#> [208] "CRE(-ROB,-RUB)(R)" "mALT(R)"           "CRE(L)"           
#> [211] "g5(R)"             "PLP(R)"            "EPA(R)"           
#> [214] "AL(R)"             "AL-VA3(R)"         "AL-VP3(R)"        
#> [217] "SMP(L)"            "LAL(-GA)(R)"       "LX(R)"            
#> [220] "AL-VM7v(R)"        "FBl9"              "AL-DM6(R)"        
#> [223] "AL-VA4(R)"         "AL-VA6(R)"         "AL-VM7d(R)"       
#> [226] "PB(L9)"            "GOR(R)"            "AL-DC3(R)"        
#> [229] "AL-VM3(R)"         "POC"              
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method     path                               
#> [1,] "PROPFIND" "/api/*"                           
#> [2,] "GET"      "/api/v:ver/npexplorer/findneurons"
#> [3,] "GET"      "/api/npexplorer/neuronmeta"       
#> [4,] "REPORT"   "/api"                             
#> [5,] "TRACE"    "/api/*"                           
#> [6,] "POST"     "/api/roimeshes/mesh/:dataset/:roi"
#>      name                                                                                      
#> [1,] "github.com/labstack/echo.glob..func1"                                                    
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                      
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMeta-fm"
#> [4,] "github.com/labstack/echo.glob..func1"                                                    
#> [5,] "github.com/labstack/echo.glob..func1"                                                    
#> [6,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                  

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }