Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "a'L(L)"            "vACA(R)"           "GA(R)"            
#>   [4] "AL-VM7d(L)"        "AL-VC5(R)"         "a2(R)"            
#>   [7] "SCL(R)"            "AL-VM7v(L)"        "CRE(L)"           
#>  [10] "AL-DM4(R)"         "AL-DM5(R)"         "AL-DP1l(R)"       
#>  [13] "AB(L)"             "AL-DM4(L)"         "LH(R)"            
#>  [16] "SPS(R)"            "WED(R)"            "FB"               
#>  [19] "a'L(R)"            "AL-DM6(R)"         "b2(R)"            
#>  [22] "MB(+ACA)(R)"       "PB(L2)"            "a'2(R)"           
#>  [25] "AL-D(L)"           "VES(R)"            "mALT(L)"          
#>  [28] "CRE(R)"            "EBr5"              "PB(R8)"           
#>  [31] "EBr6"              "SPS(L)"            "VES(L)"           
#>  [34] "AL-VC3m(R)"        "ATL(L)"            "b'2(R)"           
#>  [37] "MB(L)"             "PB(L1)"            "SNP(R)"           
#>  [40] "FBl3"              "bL(R)"             "PB(R2)"           
#>  [43] "AL-VM7v(R)"        "AL-DM2(L)"         "AVLP(R)"          
#>  [46] "aL(L)"             "CRE(-ROB,-RUB)(R)" "EBr2r4"           
#>  [49] "AL-DL5(L)"         "AL-DM3(L)"         "AL-VA2(R)"        
#>  [52] "CRE(-RUB)(L)"      "FB-column3"        "AL-DC2(R)"        
#>  [55] "AL-VP1l(R)"        "AL-VP1d(R)"        "PED(R)"           
#>  [58] "PB(L8)"            "NO(L)"             "AL-VC3l(R)"       
#>  [61] "AL-VP3(R)"         "AL-VM5v(R)"        "gL(L)"            
#>  [64] "g4(R)"             "a1(R)"             "CA(L)"            
#>  [67] "LAL(L)"            "AL-DM1(L)"         "FBl2"             
#>  [70] "AL-VL2a(R)"        "AL-DC2(L)"         "NO"               
#>  [73] "GOR(R)"            "AL-DA4l(R)"        "g3(R)"            
#>  [76] "FBl9"              "NO2(R)"            "AL-VL1(R)"        
#>  [79] "EBr1"              "SAD(-AMMC)"        "AB(R)"            
#>  [82] "AL-DL5(R)"         "AL-D(R)"           "GOR(L)"           
#>  [85] "EB"                "AL-VA3(R)"         "BU(R)"            
#>  [88] "AL-DL4(L)"         "AL-DP1m(L)"        "AL-DL4(R)"        
#>  [91] "AL-VC2(R)"         "b'L(R)"            "g2(R)"            
#>  [94] "dACA(R)"           "PB(R5)"            "AL-DL1(R)"        
#>  [97] "AL-VM4(R)"         "AL-VM7d(R)"        "IB"               
#> [100] "AOTU(R)"           "b'L(L)"            "AL-VM2(R)"        
#> [103] "SIP(R)"            "FBl1"              "b1(R)"            
#> [106] "SLP(R)"            "FLA(R)"            "FBl5"             
#> [109] "AL-DA2(R)"         "AL-VA1d(R)"        "AL-VA5(R)"        
#> [112] "AL-VA4(R)"         "AL-DA4m(L)"        "AL-VA6(L)"        
#> [115] "ATL(R)"            "EPA(L)"            "AL-DL2v(R)"       
#> [118] "AL-VA1v(R)"        "ME(R)"             "AL-DC1(R)"        
#> [121] "AL-VC4(R)"         "VLNP(R)"           "GF(R)"            
#> [124] "gL(R)"             "AL-V(R)"           "AL-VP5(R)"        
#> [127] "SAD"               "PRW"               "PB(R7)"           
#> [130] "NO3(R)"            "bL(L)"             "SNP(L)"           
#> [133] "AL-DC1(L)"         "PB(L6)"            "FBl4"             
#> [136] "AL-DL3(R)"         "AL-VP1m(R)"        "AMMC"             
#> [139] "g5(R)"             "PB(L5)"            "IB112"            
#> [142] "RUB(L)"            "SIP(L)"            "EPA(R)"           
#> [145] "a'1(R)"            "NO3(L)"            "AL-VP4(R)"        
#> [148] "AL-DC3(R)"         "AL-VA7m(R)"        "PB(R6)"           
#> [151] "SMP(L)"            "g1(R)"             "NO1(L)"           
#> [154] "NO1(R)"            "LX(R)"             "AL(R)"            
#> [157] "ROB(R)"            "INP"               "EBr3pw"           
#> [160] "FBl6"              "PB(L9)"            "FBl8"             
#> [163] "BU(L)"             "PB"                "PVLP(R)"          
#> [166] "AL-DL2d(R)"        "AL-DM2(R)"         "b'1(R)"           
#> [169] "hemibrain"         "AL-DA3(L)"         "PB(L7)"           
#> [172] "PB(R3)"            "SMP(R)"            "AL-DM5(L)"        
#> [175] "FBl7"              "a3(R)"             "GC"               
#> [178] "CAN(R)"            "AL-DA3(R)"         "AL-VA7l(R)"       
#> [181] "AL-VA6(R)"         "AOT(R)"            "POC"              
#> [184] "AL-DC4(L)"         "LO(R)"             "ICL(R)"           
#> [187] "IPS(R)"            "AL-DM3(R)"         "AL-VM1(R)"        
#> [190] "AL-VM5d(R)"        "VMNP"              "AL-VP2(R)"        
#> [193] "LOP(R)"            "LAL(-GA)(R)"       "PB(R9)"           
#> [196] "AL-DA1(R)"         "AL-DC4(R)"         "PENP"             
#> [199] "NO(R)"             "CA(R)"             "GNG"              
#> [202] "LAL(R)"            "MB(R)"             "AL-DP1m(R)"       
#> [205] "mALT(R)"           "AL-DA2(L)"         "PB(R1)"           
#> [208] "PB(L4)"            "AL-VM3(R)"         "EBr3am"           
#> [211] "lACA(R)"           "PLP(R)"            "OL(R)"            
#> [214] "AL-VL2p(R)"        "PB(L3)"            "AL-DA4m(R)"       
#> [217] "RUB(R)"            "CX"                "AL-DM1(R)"        
#> [220] "AL-VC1(R)"         "a'3(R)"            "AME(R)"           
#> [223] "PB(R4)"            "SCL(L)"            "ICL(L)"           
#> [226] "aL(R)"             "AL-DM6(L)"         "EBr3d"            
#> [229] "AL(L)"             "LX(L)"             "NO2(L)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "SPS(L)"            "CAN(R)"            "GOR(R)"           
#>   [4] "LAL(R)"            "CA(L)"             "a2(R)"            
#>   [7] "lACA(R)"           "LAL(-GA)(R)"       "NO2(L)"           
#>  [10] "FBl8"              "PENP"              "MB(L)"            
#>  [13] "SIP(R)"            "vACA(R)"           "PRW"              
#>  [16] "g5(R)"             "FBl6"              "PB(R3)"           
#>  [19] "AL(R)"             "LOP(R)"            "EBr5"             
#>  [22] "a'3(R)"            "PB(L5)"            "g3(R)"            
#>  [25] "WED(R)"            "g2(R)"             "RUB(R)"           
#>  [28] "b2(R)"             "BU(L)"             "PB(L1)"           
#>  [31] "FBl3"              "ICL(R)"            "bL(R)"            
#>  [34] "SMP(L)"            "BU(R)"             "EPA(R)"           
#>  [37] "PB(L7)"            "INP"               "PB(L4)"           
#>  [40] "CRE(-ROB,-RUB)(R)" "ATL(L)"            "SPS(R)"           
#>  [43] "SNP(R)"            "AL(L)"             "g1(R)"            
#>  [46] "NO3(L)"            "FBl7"              "PB(R4)"           
#>  [49] "LAL(L)"            "gL(L)"             "AVLP(R)"          
#>  [52] "FBl5"              "a'L(L)"            "b'2(R)"           
#>  [55] "SIP(L)"            "VES(R)"            "VLNP(R)"          
#>  [58] "g4(R)"             "CA(R)"             "LH(R)"            
#>  [61] "CX"                "NO1(R)"            "EBr3d"            
#>  [64] "EBr3pw"            "EPA(L)"            "AB(L)"            
#>  [67] "PB(R6)"            "PED(R)"            "PB"               
#>  [70] "FLA(R)"            "FBl2"              "LX(L)"            
#>  [73] "MB(R)"             "a1(R)"             "NO1(L)"           
#>  [76] "FB"                "dACA(R)"           "b1(R)"            
#>  [79] "AMMC"              "hemibrain"         "FBl1"             
#>  [82] "SCL(L)"            "ROB(R)"            "gL(R)"            
#>  [85] "PVLP(R)"           "PB(R2)"            "aL(L)"            
#>  [88] "NO3(R)"            "EBr6"              "PLP(R)"           
#>  [91] "RUB(L)"            "mALT(L)"           "ME(R)"            
#>  [94] "MB(+ACA)(R)"       "PB(L3)"            "AOTU(R)"          
#>  [97] "aL(R)"             "a3(R)"             "SLP(R)"           
#> [100] "PB(L2)"            "GF(R)"             "PB(R7)"           
#> [103] "a'2(R)"            "bL(L)"             "FBl4"             
#> [106] "b'1(R)"            "NO"                "CRE(L)"           
#> [109] "a'1(R)"            "a'L(R)"            "AOT(R)"           
#> [112] "PB(R5)"            "SMP(R)"            "GC"               
#> [115] "VMNP"              "EBr2r4"            "SNP(L)"           
#> [118] "NO2(R)"            "b'L(L)"            "mALT(R)"          
#> [121] "SCL(R)"            "PB(R9)"            "IB"               
#> [124] "PB(L8)"            "NO(R)"             "SAD(-AMMC)"       
#> [127] "CRE(R)"            "EBr1"              "OL(R)"            
#> [130] "PB(R8)"            "PB(L9)"            "AME(R)"           
#> [133] "IPS(R)"            "LX(R)"             "ATL(R)"           
#> [136] "AB(R)"             "NO(L)"             "VES(L)"           
#> [139] "GA(R)"             "GOR(L)"            "GNG"              
#> [142] "EB"                "POC"               "EBr3am"           
#> [145] "SAD"               "b'L(R)"            "FBl9"             
#> [148] "LO(R)"             "PB(L6)"            "PB(R1)"           
#> [151] "ICL(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "g4(R)"             "MB(L)"             "b'L(R)"           
#>   [4] "AL-VA1v(R)"        "AL(L)"             "FB-column3"       
#>   [7] "AL-DM6(R)"         "NO1(L)"            "PB(R8)"           
#>  [10] "a'2(R)"            "AL-DC4(L)"         "AL-DC2(R)"        
#>  [13] "GC"                "PB(L9)"            "SNP(L)"           
#>  [16] "AME(R)"            "AL-VA6(L)"         "RUB(L)"           
#>  [19] "AL-DP1l(R)"        "CX"                "PRW"              
#>  [22] "AL-DM2(L)"         "FBl8"              "PB(R3)"           
#>  [25] "CAN(R)"            "PB(L3)"            "AB(R)"            
#>  [28] "PB(L8)"            "NO"                "MB(+ACA)(R)"      
#>  [31] "OL(R)"             "FBl4"              "AL-DL2v(R)"       
#>  [34] "VLNP(R)"           "NO(R)"             "AL-DL2d(R)"       
#>  [37] "AL-VM7d(L)"        "ATL(R)"            "PB(R4)"           
#>  [40] "AL-VC4(R)"         "AL-VM4(R)"         "a'1(R)"           
#>  [43] "FBl7"              "VES(L)"            "AL-VP1m(R)"       
#>  [46] "LO(R)"             "AL-DL4(R)"         "AOT(R)"           
#>  [49] "mALT(R)"           "SNP(R)"            "PVLP(R)"          
#>  [52] "LX(L)"             "EBr2r4"            "CRE(-RUB)(L)"     
#>  [55] "FBl3"              "GOR(R)"            "PB(L4)"           
#>  [58] "AL-DC3(R)"         "AL-VP3(R)"         "LX(R)"            
#>  [61] "gL(L)"             "aL(L)"             "AL-DM1(R)"        
#>  [64] "AL-VL2a(R)"        "AL-VA2(R)"         "LOP(R)"           
#>  [67] "PENP"              "AL-DC1(L)"         "FBl5"             
#>  [70] "AL-VP2(R)"         "CA(L)"             "NO3(R)"           
#>  [73] "PB(R1)"            "bL(R)"             "AL-DM2(R)"        
#>  [76] "IB"                "PB(L5)"            "SPS(R)"           
#>  [79] "ATL(L)"            "AL-DP1m(L)"        "AVLP(R)"          
#>  [82] "b'L(L)"            "RUB(R)"            "a3(R)"            
#>  [85] "AL-DM3(R)"         "AL-VC5(R)"         "WED(R)"           
#>  [88] "SPS(L)"            "AL-DM1(L)"         "SIP(R)"           
#>  [91] "AL-DA1(R)"         "CRE(-ROB,-RUB)(R)" "AL-VA4(R)"        
#>  [94] "FBl6"              "PB(R6)"            "NO2(R)"           
#>  [97] "a1(R)"             "a2(R)"             "ROB(R)"           
#> [100] "AL-V(R)"           "lACA(R)"           "g5(R)"            
#> [103] "FBl2"              "AL-VP4(R)"         "a'3(R)"           
#> [106] "vACA(R)"           "PED(R)"            "PB(L1)"           
#> [109] "PB(L6)"            "AL-DA2(R)"         "g1(R)"            
#> [112] "aL(R)"             "VMNP"              "SIP(L)"           
#> [115] "AL-VM7v(L)"        "gL(R)"             "AL-DC1(R)"        
#> [118] "AL-DC4(R)"         "AL-DL5(R)"         "AL-VP5(R)"        
#> [121] "FB"                "AL-VC2(R)"         "SCL(R)"           
#> [124] "AL-DA2(L)"         "EBr3pw"            "PB(L7)"           
#> [127] "IPS(R)"            "AL-DL1(R)"         "AL-VM5v(R)"       
#> [130] "AL-VM7v(R)"        "AL-DL5(L)"         "FBl1"             
#> [133] "AMMC"              "AL-DL4(L)"         "bL(L)"            
#> [136] "ICL(R)"            "g3(R)"             "AL-DM4(L)"        
#> [139] "AL-DM5(L)"         "NO1(R)"            "EPA(R)"           
#> [142] "AL-D(L)"           "NO3(L)"            "FBl9"             
#> [145] "CRE(L)"            "VES(R)"            "SCL(L)"           
#> [148] "NO2(L)"            "EBr3d"             "AL-VM5d(R)"       
#> [151] "AL-VP1l(R)"        "b'2(R)"            "PLP(R)"           
#> [154] "AL-DA4l(R)"        "AL-VL1(R)"         "AL-DA4m(R)"       
#> [157] "AL-VL2p(R)"        "POC"               "b'1(R)"           
#> [160] "EB"                "GNG"               "AL-VC1(R)"        
#> [163] "AL-DM3(L)"         "PB(R9)"            "ICL(L)"           
#> [166] "PB(R2)"            "AL-DA3(R)"         "AL-VA7l(R)"       
#> [169] "AL-VM1(R)"         "AL-VP1d(R)"        "EPA(L)"           
#> [172] "AL-DP1m(R)"        "SMP(L)"            "GA(R)"            
#> [175] "b2(R)"             "a'L(L)"            "AL-DM4(R)"        
#> [178] "AL-VA6(R)"         "SMP(R)"            "EBr3am"           
#> [181] "g2(R)"             "LH(R)"             "b1(R)"            
#> [184] "AL-DL3(R)"         "AL-VA7m(R)"        "AOTU(R)"          
#> [187] "CRE(R)"            "AL-DM6(L)"         "PB(L2)"           
#> [190] "PB(R7)"            "GOR(L)"            "a'L(R)"           
#> [193] "LAL(R)"            "hemibrain"         "LAL(-GA)(R)"      
#> [196] "AL-DM5(R)"         "AL-VM3(R)"         "PB(R5)"           
#> [199] "EBr6"              "dACA(R)"           "AL-VA3(R)"        
#> [202] "SAD"               "BU(R)"             "AL-VA5(R)"        
#> [205] "AL-VC3m(R)"        "PB"                "ME(R)"            
#> [208] "AL-VA1d(R)"        "GF(R)"             "BU(L)"            
#> [211] "AL-DA4m(L)"        "INP"               "FLA(R)"           
#> [214] "AL-DC2(L)"         "SAD(-AMMC)"        "SLP(R)"           
#> [217] "MB(R)"             "CA(R)"             "NO(L)"            
#> [220] "AL(R)"             "mALT(L)"           "LAL(L)"           
#> [223] "EBr1"              "EBr5"              "AL-D(R)"          
#> [226] "AL-VM2(R)"         "AL-DA3(L)"         "AL-VC3l(R)"       
#> [229] "AL-VM7d(R)"        "AB(L)"            
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "a1(R)"             "AL-VC5(R)"         "AL-VM5d(R)"       
#>   [4] "AL(L)"             "EBr3am"            "MB(L)"            
#>   [7] "AL-VM7v(L)"        "g5(R)"             "PB(R3)"           
#>  [10] "ATL(L)"            "a'L(R)"            "AL-DC4(R)"        
#>  [13] "vACA(R)"           "NO1(R)"            "SCL(L)"           
#>  [16] "PB(L2)"            "AL-VA4(R)"         "IPS(R)"           
#>  [19] "AL-DA1(R)"         "AL-VL1(R)"         "SAD"              
#>  [22] "AL-VA2(R)"         "AL-VP3(R)"         "PB(R5)"           
#>  [25] "SAD(-AMMC)"        "INP"               "OL(R)"            
#>  [28] "NO(R)"             "GNG"               "AL-D(R)"          
#>  [31] "AL-VP5(R)"         "lACA(R)"           "AL-DL4(L)"        
#>  [34] "SIP(R)"            "bL(R)"             "LAL(R)"           
#>  [37] "AL-DA3(L)"         "PENP"              "PB(L9)"           
#>  [40] "FBl7"              "AL-VM3(R)"         "NO1(L)"           
#>  [43] "VMNP"              "EBr3d"             "AL-DC4(L)"        
#>  [46] "FBl4"              "AL-DA4l(R)"        "AL-VA7l(R)"       
#>  [49] "AL-VP4(R)"         "PB"                "AL-VP1m(R)"       
#>  [52] "AL-DM2(R)"         "SPS(L)"            "AL-VC3m(R)"       
#>  [55] "PB(R6)"            "LAL(-GA)(R)"       "AL-DA2(L)"        
#>  [58] "mALT(L)"           "PLP(R)"            "EBr2r4"           
#>  [61] "AL-DM2(L)"         "AL-DP1m(L)"        "AL-DL3(R)"        
#>  [64] "AL-VC4(R)"         "AL-VM2(R)"         "b2(R)"            
#>  [67] "MB(+ACA)(R)"       "FBl2"              "VES(L)"           
#>  [70] "AL-VP1d(R)"        "GA(R)"             "a2(R)"            
#>  [73] "PB(L1)"            "bL(L)"             "AL-VM5v(R)"       
#>  [76] "LOP(R)"            "PB(L5)"            "GF(R)"            
#>  [79] "SNP(L)"            "NO3(R)"            "LX(R)"            
#>  [82] "ROB(R)"            "LH(R)"             "AL-DM4(R)"        
#>  [85] "AL-VA6(R)"         "SMP(L)"            "AL-DM5(L)"        
#>  [88] "PB(R1)"            "AL(R)"             "AL-DA4m(R)"       
#>  [91] "BU(L)"             "FBl8"              "AL-VC3l(R)"       
#>  [94] "AL-VP1l(R)"        "SMP(R)"            "AL-VA6(L)"        
#>  [97] "SNP(R)"            "PB(L6)"            "AL-VA7m(R)"       
#> [100] "VLNP(R)"           "a'3(R)"            "SPS(R)"           
#> [103] "AL-DM5(R)"         "PB(L4)"            "CRE(-RUB)(L)"     
#> [106] "PVLP(R)"           "gL(R)"             "PB(R9)"           
#> [109] "RUB(L)"            "AL-DM1(R)"         "AL-DP1l(R)"       
#> [112] "hemibrain"         "LO(R)"             "CA(R)"            
#> [115] "ICL(R)"            "a3(R)"             "AL-DA3(R)"        
#> [118] "AB(L)"             "PB(L3)"            "PB(R4)"           
#> [121] "FB-column3"        "PB(L8)"            "dACA(R)"          
#> [124] "EPA(L)"            "AL-DL1(R)"         "CA(L)"            
#> [127] "GC"                "AMMC"              "AL-DM4(L)"        
#> [130] "b'L(L)"            "AL-DC2(R)"         "b'1(R)"           
#> [133] "CX"                "EPA(R)"            "EB"               
#> [136] "ATL(R)"            "AL-DA2(R)"         "g2(R)"            
#> [139] "AL-VA1v(R)"        "AL-VC1(R)"         "AL-VM7v(R)"       
#> [142] "POC"               "AOTU(R)"           "CRE(R)"           
#> [145] "EBr5"              "WED(R)"            "AL-VA1d(R)"       
#> [148] "AL-VC2(R)"         "BU(R)"             "PB(L7)"           
#> [151] "LAL(L)"            "FLA(R)"            "AL-VM1(R)"        
#> [154] "EBr6"              "a'1(R)"            "aL(L)"            
#> [157] "AL-DL2v(R)"        "RUB(R)"            "PED(R)"           
#> [160] "NO"                "AL-DL4(R)"         "AL-DL5(R)"        
#> [163] "AL-VM7d(L)"        "FBl3"              "LX(L)"            
#> [166] "AL-DM3(R)"         "PB(R8)"            "AL-VA5(R)"        
#> [169] "ME(R)"             "EBr1"              "FBl6"             
#> [172] "PB(R2)"            "CAN(R)"            "AL-V(R)"          
#> [175] "g1(R)"             "AL-DM3(L)"         "VES(R)"           
#> [178] "GOR(R)"            "AL-VM7d(R)"        "PB(R7)"           
#> [181] "AME(R)"            "NO3(L)"            "AVLP(R)"          
#> [184] "AL-DL2d(R)"        "AL-VP2(R)"         "AL-DC1(L)"        
#> [187] "AL-DC3(R)"         "SCL(R)"            "mALT(R)"          
#> [190] "AL-DM6(L)"         "FBl1"              "AL-VA3(R)"        
#> [193] "CRE(-ROB,-RUB)(R)" "b'2(R)"            "AL-DM1(L)"        
#> [196] "PRW"               "g4(R)"             "AL-DL5(L)"        
#> [199] "b1(R)"             "MB(R)"             "AOT(R)"           
#> [202] "g3(R)"             "CRE(L)"            "ICL(L)"           
#> [205] "AB(R)"             "NO(L)"             "IB"               
#> [208] "FB"                "AL-VL2a(R)"        "SIP(L)"           
#> [211] "FBl9"              "AL-DC1(R)"         "AL-DM6(R)"        
#> [214] "AL-DP1m(R)"        "AL-D(L)"           "b'L(R)"           
#> [217] "AL-DC2(L)"         "EBr3pw"            "a'2(R)"           
#> [220] "gL(L)"             "a'L(L)"            "AL-VL2p(R)"       
#> [223] "AL-VM4(R)"         "NO2(R)"            "aL(R)"            
#> [226] "AL-DA4m(L)"        "NO2(L)"            "GOR(L)"           
#> [229] "SLP(R)"            "FBl5"             
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                        
#> [1,] "PUT"  "/api"                                      
#> [2,] "POST" "/api/v:ver/npexplorer/rankedtable"         
#> [3,] "POST" "/api/v:ver/npexplorer/roisinneuron"        
#> [4,] "GET"  "/api/npexplorer/completeness"              
#> [5,] "GET"  "/api/v:ver/cached/roicompleteness"         
#> [6,] "POST" "/api/v:ver/raw/keyvalue/key/:instance/:key"
#>      name                                                                                        
#> [1,] "github.com/labstack/echo.glob..func1"                                                      
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCompleteness-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }