Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "NO(L)"             "LOP(R)"            "g2(R)"            
#>   [4] "a'1(R)"            "GOR(L)"            "PB(L1)"           
#>   [7] "ME(R)"             "CRE(L)"            "lACA(R)"          
#>  [10] "PB(R7)"            "IPS(R)"            "LAL(L)"           
#>  [13] "GC"                "WED(R)"            "dACA(R)"          
#>  [16] "ROB(R)"            "FBl5"              "RUB(L)"           
#>  [19] "BU(L)"             "SIP(R)"            "NO2(R)"           
#>  [22] "a3(R)"             "IB"                "FB"               
#>  [25] "a'2(R)"            "EB"                "GNG"              
#>  [28] "PB(L8)"            "LX(R)"             "EBr2r4"           
#>  [31] "NO2(L)"            "GF(R)"             "PED(R)"           
#>  [34] "b2(R)"             "NO1(R)"            "PLP(R)"           
#>  [37] "GOR(R)"            "b'1(R)"            "PB(L2)"           
#>  [40] "a2(R)"             "mALT(R)"           "AB(R)"            
#>  [43] "OL(R)"             "FBl3"              "FBl4"             
#>  [46] "EBr3am"            "CX"                "VES(R)"           
#>  [49] "SCL(L)"            "gL(L)"             "NO3(L)"           
#>  [52] "SNP(L)"            "VES(L)"            "a1(R)"            
#>  [55] "NO3(R)"            "CRE(R)"            "SCL(R)"           
#>  [58] "SPS(L)"            "EBr3d"             "a'L(R)"           
#>  [61] "ICL(L)"            "PB(L7)"            "g3(R)"            
#>  [64] "MB(L)"             "CAN(R)"            "MB(R)"            
#>  [67] "SNP(R)"            "MB(+ACA)(R)"       "VMNP"             
#>  [70] "PB(L6)"            "PVLP(R)"           "CA(R)"            
#>  [73] "AOT(R)"            "RUB(R)"            "SAD"              
#>  [76] "FLA(R)"            "a'3(R)"            "vACA(R)"          
#>  [79] "PB(L9)"            "LX(L)"             "LAL(R)"           
#>  [82] "EBr6"              "EBr5"              "b'L(R)"           
#>  [85] "EPA(R)"            "AME(R)"            "PB(L4)"           
#>  [88] "g4(R)"             "PB(R3)"            "NO1(L)"           
#>  [91] "PB(R8)"            "EBr3pw"            "FBl9"             
#>  [94] "PB(R2)"            "ATL(R)"            "b'L(L)"           
#>  [97] "g1(R)"             "aL(R)"             "ATL(L)"           
#> [100] "FBl8"              "PB(L5)"            "AOTU(R)"          
#> [103] "SAD(-AMMC)"        "LO(R)"             "b1(R)"            
#> [106] "PB(R1)"            "bL(L)"             "FBl1"             
#> [109] "AMMC"              "SIP(L)"            "PB(R9)"           
#> [112] "NO"                "gL(R)"             "FBl7"             
#> [115] "bL(R)"             "LAL(-GA)(R)"       "EBr1"             
#> [118] "aL(L)"             "EPA(L)"            "SLP(R)"           
#> [121] "VLNP(R)"           "POC"               "AB(L)"            
#> [124] "PB(R6)"            "GA(R)"             "PRW"              
#> [127] "g5(R)"             "FBl6"              "AVLP(R)"          
#> [130] "mALT(L)"           "SMP(L)"            "hemibrain"        
#> [133] "CRE(-ROB,-RUB)(R)" "BU(R)"             "INP"              
#> [136] "PB(R4)"            "a'L(L)"            "b'2(R)"           
#> [139] "SPS(R)"            "PB(R5)"            "PB"               
#> [142] "AL(L)"             "NO(R)"             "FBl2"             
#> [145] "LH(R)"             "ICL(R)"            "PENP"             
#> [148] "PB(L3)"            "CA(L)"             "AL(R)"            
#> [151] "SMP(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "gL(L)"             "AL-DA3(L)"         "AL-DM2(L)"        
#>   [4] "ICL(L)"            "AL-DA3(R)"         "a2(R)"            
#>   [7] "GOR(L)"            "AL-DA4m(R)"        "EBr3am"           
#>  [10] "PENP"              "FLA(R)"            "a'L(R)"           
#>  [13] "PB(R1)"            "PVLP(R)"           "NO(L)"            
#>  [16] "AL-DL2v(R)"        "AL-DP1l(R)"        "PRW"              
#>  [19] "LAL(-GA)(R)"       "AL-VP1l(R)"        "LH(R)"            
#>  [22] "AVLP(R)"           "PB(L2)"            "FBl3"             
#>  [25] "AL-VC3m(R)"        "PB(L6)"            "CA(R)"            
#>  [28] "RUB(R)"            "vACA(R)"           "LAL(L)"           
#>  [31] "NO1(R)"            "SNP(L)"            "AL-VL2a(R)"       
#>  [34] "AL-D(R)"           "AL-VA3(R)"         "AL-V(R)"          
#>  [37] "PB(R8)"            "MB(R)"             "VES(R)"           
#>  [40] "MB(L)"             "g5(R)"             "AL-VP2(R)"        
#>  [43] "CRE(L)"            "AL-VL1(R)"         "AL-VL2p(R)"       
#>  [46] "PB(L1)"            "FBl1"              "SMP(L)"           
#>  [49] "VLNP(R)"           "AL-VM7d(R)"        "AL-VP5(R)"        
#>  [52] "AL-VA6(L)"         "AL-VA2(R)"         "IPS(R)"           
#>  [55] "LX(L)"             "AL-VC1(R)"         "b'2(R)"           
#>  [58] "WED(R)"            "SPS(R)"            "g1(R)"            
#>  [61] "lACA(R)"           "AL-VM5d(R)"        "FBl5"             
#>  [64] "b'L(L)"            "LX(R)"             "EBr3d"            
#>  [67] "AL-DM4(L)"         "g3(R)"             "SCL(R)"           
#>  [70] "SAD(-AMMC)"        "dACA(R)"           "ROB(R)"           
#>  [73] "AMMC"              "CRE(-RUB)(L)"      "AL-VP3(R)"        
#>  [76] "FBl8"              "FBl9"              "POC"              
#>  [79] "a'3(R)"            "AL-DA2(L)"         "MB(+ACA)(R)"      
#>  [82] "AB(L)"             "AL(R)"             "AL-VM1(R)"        
#>  [85] "EBr6"              "AL-DP1m(R)"        "GF(R)"            
#>  [88] "GA(R)"             "AL-VC4(R)"         "LOP(R)"           
#>  [91] "hemibrain"         "PB(R9)"            "AL-VM7v(R)"       
#>  [94] "aL(R)"             "PLP(R)"            "AL-VP1d(R)"       
#>  [97] "EPA(R)"            "LO(R)"             "NO(R)"            
#> [100] "EPA(L)"            "SMP(R)"            "b2(R)"            
#> [103] "EB"                "AL-DC2(L)"         "IB"               
#> [106] "PB(L3)"            "EBr1"              "AOT(R)"           
#> [109] "CRE(R)"            "AL-DC3(R)"         "AL-DM6(R)"        
#> [112] "NO2(L)"            "a'2(R)"            "AL-DA4m(L)"       
#> [115] "SLP(R)"            "AL-DM5(R)"         "PB(R7)"           
#> [118] "g2(R)"             "a'L(L)"            "OL(R)"            
#> [121] "VMNP"              "SIP(R)"            "PB"               
#> [124] "AL-VA4(R)"         "AL-DM6(L)"         "AL-D(L)"          
#> [127] "AL-DM3(L)"         "RUB(L)"            "AL-VM2(R)"        
#> [130] "GC"                "AL-DL5(L)"         "FB-column3"       
#> [133] "AL-VA1d(R)"        "AME(R)"            "VES(L)"           
#> [136] "AL-VM4(R)"         "AL-VM7v(L)"        "LAL(R)"           
#> [139] "AL-DA1(R)"         "AL-VA7m(R)"        "b'L(R)"           
#> [142] "SNP(R)"            "BU(R)"             "AL-DC1(L)"        
#> [145] "a'1(R)"            "FBl7"              "PB(L4)"           
#> [148] "AL-DL5(R)"         "EBr5"              "ME(R)"            
#> [151] "SCL(L)"            "CRE(-ROB,-RUB)(R)" "AL-VM3(R)"        
#> [154] "AL-VP4(R)"         "AL-VP1m(R)"        "PB(R5)"           
#> [157] "AOTU(R)"           "EBr2r4"            "bL(L)"            
#> [160] "AL-VC5(R)"         "PB(R2)"            "AL-VA6(R)"        
#> [163] "ICL(R)"            "AL-VC3l(R)"        "AL-DC2(R)"        
#> [166] "PB(L5)"            "mALT(R)"           "EBr3pw"           
#> [169] "NO"                "AL-DA4l(R)"        "AL-VC2(R)"        
#> [172] "CX"                "PB(L8)"            "AL(L)"            
#> [175] "gL(R)"             "ATL(L)"            "AL-DC4(L)"        
#> [178] "AL-VM7d(L)"        "GNG"               "AL-DA2(R)"        
#> [181] "AL-DP1m(L)"        "b1(R)"             "AL-DM4(R)"        
#> [184] "b'1(R)"            "AL-DM5(L)"         "AB(R)"            
#> [187] "AL-DC1(R)"         "PB(R6)"            "SPS(L)"           
#> [190] "PB(R4)"            "FBl6"              "NO2(R)"           
#> [193] "AL-DL4(R)"         "AL-VA7l(R)"        "BU(L)"            
#> [196] "a3(R)"             "CAN(R)"            "aL(L)"            
#> [199] "AL-DL2d(R)"        "AL-DM2(R)"         "PED(R)"           
#> [202] "ATL(R)"            "FB"                "AL-DL4(L)"        
#> [205] "AL-DL1(R)"         "AL-DM1(R)"         "SAD"              
#> [208] "g4(R)"             "AL-VA5(R)"         "NO3(R)"           
#> [211] "PB(L9)"            "AL-DL3(R)"         "CA(L)"            
#> [214] "AL-VA1v(R)"        "AL-DM1(L)"         "PB(L7)"           
#> [217] "GOR(R)"            "AL-DC4(R)"         "AL-DM3(R)"        
#> [220] "SIP(L)"            "FBl4"              "PB(R3)"           
#> [223] "bL(R)"             "AL-VM5v(R)"        "mALT(L)"          
#> [226] "NO1(L)"            "FBl2"              "NO3(L)"           
#> [229] "a1(R)"             "INP"              
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method     path                                     
#> [1,] "POST"     "/api"                                   
#> [2,] "PROPFIND" "/api"                                   
#> [3,] "GET"      "/api/v:ver/cached/roicompleteness"      
#> [4,] "POST"     "/api/v:ver/roimeshes/mesh/:dataset/:roi"
#> [5,] "GET"      "/api/v:ver/dbmeta/version"              
#> [6,] "CONNECT"  "/api/*"                                 
#>      name                                                                
#> [1,] "github.com/labstack/echo.glob..func1"                              
#> [2,] "github.com/labstack/echo.glob..func1"                              
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [6,] "github.com/labstack/echo.glob..func1"                              

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }