Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "g3(R)"             "GOR(L)"            "LX(L)"            
#>   [4] "NO2(L)"            "EBr6"              "NO1(R)"           
#>   [7] "PB(R8)"            "EBr3pw"            "FBl3"             
#>  [10] "LH(R)"             "b'L(L)"            "PB(L4)"           
#>  [13] "VLNP(R)"           "AOT(R)"            "hemibrain"        
#>  [16] "PB(L7)"            "PB(R2)"            "PB(R3)"           
#>  [19] "a3(R)"             "SMP(R)"            "mALT(L)"          
#>  [22] "SPS(R)"            "PED(R)"            "a'2(R)"           
#>  [25] "GC"                "ATL(R)"            "FBl4"             
#>  [28] "aL(L)"             "AL(R)"             "NO1(L)"           
#>  [31] "PB(R9)"            "RUB(L)"            "SCL(L)"           
#>  [34] "EBr3d"             "LO(R)"             "CAN(R)"           
#>  [37] "MB(L)"             "gL(L)"             "b'1(R)"           
#>  [40] "BU(L)"             "SNP(L)"            "AME(R)"           
#>  [43] "POC"               "mALT(R)"           "ROB(R)"           
#>  [46] "VES(L)"            "PB(R7)"            "PVLP(R)"          
#>  [49] "MB(+ACA)(R)"       "bL(R)"             "CRE(-ROB,-RUB)(R)"
#>  [52] "PB(R6)"            "NO(L)"             "LOP(R)"           
#>  [55] "AMMC"              "SAD"               "aL(R)"            
#>  [58] "FB"                "a'L(R)"            "g4(R)"            
#>  [61] "a'L(L)"            "PB(R5)"            "CRE(L)"           
#>  [64] "FBl6"              "SLP(R)"            "GF(R)"            
#>  [67] "GA(R)"             "b'L(R)"            "PB(L6)"           
#>  [70] "GOR(R)"            "AB(L)"             "IB"               
#>  [73] "CA(L)"             "a2(R)"             "SIP(L)"           
#>  [76] "g1(R)"             "SCL(R)"            "PB"               
#>  [79] "vACA(R)"           "LAL(L)"            "WED(R)"           
#>  [82] "EB"                "AL(L)"             "LAL(R)"           
#>  [85] "RUB(R)"            "a'3(R)"            "g2(R)"            
#>  [88] "EBr5"              "FLA(R)"            "gL(R)"            
#>  [91] "a'1(R)"            "FBl1"              "b1(R)"            
#>  [94] "PENP"              "dACA(R)"           "VMNP"             
#>  [97] "g5(R)"             "IPS(R)"            "NO2(R)"           
#> [100] "CX"                "EBr1"              "ME(R)"            
#> [103] "SIP(R)"            "NO3(L)"            "PB(L2)"           
#> [106] "FBl5"              "FBl8"              "INP"              
#> [109] "ICL(L)"            "FBl2"              "EPA(L)"           
#> [112] "PB(L3)"            "PB(L8)"            "NO(R)"            
#> [115] "ICL(R)"            "NO3(R)"            "CRE(R)"           
#> [118] "SAD(-AMMC)"        "MB(R)"             "LX(R)"            
#> [121] "FBl9"              "OL(R)"             "lACA(R)"          
#> [124] "PLP(R)"            "NO"                "PB(R1)"           
#> [127] "VES(R)"            "CA(R)"             "GNG"              
#> [130] "b'2(R)"            "PRW"               "PB(L1)"           
#> [133] "FBl7"              "AB(R)"             "PB(R4)"           
#> [136] "PB(L5)"            "SMP(L)"            "AOTU(R)"          
#> [139] "AVLP(R)"           "ATL(L)"            "SPS(L)"           
#> [142] "EBr2r4"            "a1(R)"             "b2(R)"            
#> [145] "BU(R)"             "EBr3am"            "LAL(-GA)(R)"      
#> [148] "SNP(R)"            "EPA(R)"            "PB(L9)"           
#> [151] "bL(L)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "SNP(R)"            "SAD(-AMMC)"        "MB(+ACA)(R)"      
#>   [4] "FBl1"              "SAD"               "PRW"              
#>   [7] "NO1(L)"            "EBr3d"             "SCL(R)"           
#>  [10] "PB(R3)"            "lACA(R)"           "PLP(R)"           
#>  [13] "AMMC"              "AL-DM5(L)"         "PB(R4)"           
#>  [16] "GF(R)"             "PB(R7)"            "GC"               
#>  [19] "dACA(R)"           "a2(R)"             "VMNP"             
#>  [22] "FBl8"              "AL-VC3l(R)"        "NO2(L)"           
#>  [25] "PB(L8)"            "NO2(R)"            "AL-VM3(R)"        
#>  [28] "NO3(R)"            "FB"                "ATL(R)"           
#>  [31] "AL-DA3(R)"         "AL-VC5(R)"         "PB(R6)"           
#>  [34] "FBl6"              "a'L(L)"            "AL-DC3(R)"        
#>  [37] "AL-VC2(R)"         "AL-VC3m(R)"        "AL-VP4(R)"        
#>  [40] "POC"               "SMP(L)"            "EBr1"             
#>  [43] "NO"                "RUB(L)"            "PB(L3)"           
#>  [46] "FBl7"              "vACA(R)"           "CA(L)"            
#>  [49] "AL-VP1l(R)"        "AB(L)"             "IB"               
#>  [52] "CX"                "AL(L)"             "AL-DP1m(L)"       
#>  [55] "NO(R)"             "AL-VA1v(R)"        "AL-DL5(R)"        
#>  [58] "AL-VM7v(R)"        "RUB(R)"            "AL-DM1(L)"        
#>  [61] "LO(R)"             "AL-VA4(R)"         "AL-DC4(L)"        
#>  [64] "FBl4"              "AL-VL2p(R)"        "a1(R)"            
#>  [67] "PB"                "g5(R)"             "aL(L)"            
#>  [70] "OL(R)"             "IPS(R)"            "AL-DL1(R)"        
#>  [73] "AL-VA3(R)"         "VLNP(R)"           "b2(R)"            
#>  [76] "BU(L)"             "CRE(-RUB)(L)"      "PB(L7)"           
#>  [79] "EBr3am"            "BU(R)"             "AL-DA4m(L)"       
#>  [82] "AL-DL4(R)"         "AL-DM4(R)"         "SPS(L)"           
#>  [85] "AL-DL4(L)"         "AL-VM1(R)"         "ATL(L)"           
#>  [88] "a'L(R)"            "AOT(R)"            "AL-DC4(R)"        
#>  [91] "AL-DA2(R)"         "AL-DM3(R)"         "AL-VP1m(R)"       
#>  [94] "PED(R)"            "b1(R)"             "AL-VA2(R)"        
#>  [97] "ME(R)"             "b'L(L)"            "EB"               
#> [100] "g4(R)"             "AL-DC2(L)"         "LAL(R)"           
#> [103] "AL-DA4l(R)"        "AL-VA7l(R)"        "SMP(R)"           
#> [106] "NO1(R)"            "PB(R2)"            "AME(R)"           
#> [109] "AL-DA4m(R)"        "SCL(L)"            "MB(L)"            
#> [112] "FLA(R)"            "GOR(R)"            "AL-DA1(R)"        
#> [115] "AL-VM5v(R)"        "GOR(L)"            "AL-DC1(L)"        
#> [118] "LH(R)"             "AL-VL2a(R)"        "AL-DC1(R)"        
#> [121] "AL-DP1l(R)"        "mALT(R)"           "AL-VA6(L)"        
#> [124] "FBl5"              "LX(L)"             "AL-VL1(R)"        
#> [127] "g3(R)"             "AL-DM1(R)"         "AL-D(R)"          
#> [130] "AL-VA1d(R)"        "AL-VA5(R)"         "AL-VP3(R)"        
#> [133] "FB-column3"        "SLP(R)"            "AL-DM6(R)"        
#> [136] "mALT(L)"           "EBr2r4"            "AL-DM3(L)"        
#> [139] "AVLP(R)"           "WED(R)"            "gL(L)"            
#> [142] "FBl9"              "AL-DM2(R)"         "PB(L2)"           
#> [145] "PENP"              "GA(R)"             "PB(R8)"           
#> [148] "AL-VM7d(R)"        "AL-DP1m(R)"        "NO3(L)"           
#> [151] "AB(R)"             "LAL(-GA)(R)"       "PB(L6)"           
#> [154] "SIP(R)"            "AL-DL2d(R)"        "LAL(L)"           
#> [157] "PB(L1)"            "AL-D(L)"           "AL-DM4(L)"        
#> [160] "AOTU(R)"           "AL-DA2(L)"         "AL-VM7d(L)"       
#> [163] "AL-VC4(R)"         "SIP(L)"            "bL(L)"            
#> [166] "INP"               "AL-VM5d(R)"        "PVLP(R)"          
#> [169] "hemibrain"         "aL(R)"             "AL-DM2(L)"        
#> [172] "AL-DL3(R)"         "g1(R)"             "g2(R)"            
#> [175] "AL-DM6(L)"         "FBl2"              "GNG"              
#> [178] "NO(L)"             "SPS(R)"            "EBr3pw"           
#> [181] "ICL(L)"            "FBl3"              "AL-VM7v(L)"       
#> [184] "ICL(R)"            "a3(R)"             "AL-VP2(R)"        
#> [187] "CRE(-ROB,-RUB)(R)" "b'2(R)"            "a'2(R)"           
#> [190] "gL(R)"             "AL(R)"             "AL-VM4(R)"        
#> [193] "AL-V(R)"           "PB(R5)"            "EPA(R)"           
#> [196] "a'1(R)"            "bL(R)"             "AL-VC1(R)"        
#> [199] "EBr5"              "AL-DL5(L)"         "EPA(L)"           
#> [202] "MB(R)"             "PB(L9)"            "SNP(L)"           
#> [205] "b'1(R)"            "LX(R)"             "b'L(R)"           
#> [208] "PB(L5)"            "a'3(R)"            "AL-DA3(L)"        
#> [211] "ROB(R)"            "VES(L)"            "AL-DC2(R)"        
#> [214] "CRE(R)"            "CA(R)"             "AL-VM2(R)"        
#> [217] "PB(L4)"            "AL-DM5(R)"         "EBr6"             
#> [220] "AL-VA7m(R)"        "AL-VA6(R)"         "LOP(R)"           
#> [223] "PB(R9)"            "CRE(L)"            "VES(R)"           
#> [226] "CAN(R)"            "PB(R1)"            "AL-DL2v(R)"       
#> [229] "AL-VP5(R)"         "AL-VP1d(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method  path                                        
#> [1,] "GET"   "/oauth2callback"                           
#> [2,] "GET"   "/api/v:ver/skeletons/skeleton/:dataset/:id"
#> [3,] "GET"   "/api/cached/dailytype"                     
#> [4,] "POST"  "/api/v:ver/skeletons/skeleton/:dataset/:id"
#> [5,] "POST"  "/api/v:ver/raw/keyvalue/key/:instance/:key"
#> [6,] "PATCH" "/api"                                      
#>      name                                                                              
#> [1,] "github.com/janelia-flyem/echo-secure.oauthCallbackHandler"                       
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"              
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getDailyType-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"              
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"              
#> [6,] "github.com/labstack/echo.glob..func1"                                            

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }