Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DA4m(R)"        "AL-VC1(R)"         "AL-VP1l(R)"       
#>   [4] "PB(L5)"            "FLA(R)"            "NO2(L)"           
#>   [7] "NO"                "CRE(L)"            "FBl6"             
#>  [10] "IPS(R)"            "MB(R)"             "AL-DC3(R)"        
#>  [13] "AL-VA6(R)"         "a'2(R)"            "AL-DC2(L)"        
#>  [16] "SLP(R)"            "AOT(R)"            "dACA(R)"          
#>  [19] "RUB(L)"            "b'L(L)"            "PB(L2)"           
#>  [22] "gL(R)"             "PRW"               "PB(R8)"           
#>  [25] "AL-DL4(L)"         "AL-DM2(L)"         "AL-VA3(R)"        
#>  [28] "GF(R)"             "GNG"               "FBl2"             
#>  [31] "PB(L6)"            "AL-VP1m(R)"        "AL-VM7v(L)"       
#>  [34] "AL-VM7d(L)"        "AL-DC2(R)"         "a1(R)"            
#>  [37] "LAL(-GA)(R)"       "VES(L)"            "FBl3"             
#>  [40] "PB(L3)"            "GOR(L)"            "a'3(R)"           
#>  [43] "g3(R)"             "PB(R3)"            "b1(R)"            
#>  [46] "AL-VA2(R)"         "EBr3d"             "NO(R)"            
#>  [49] "AL-VC2(R)"         "FBl5"              "AL-VM4(R)"        
#>  [52] "EBr6"              "PB(R5)"            "g1(R)"            
#>  [55] "vACA(R)"           "WED(R)"            "PENP"             
#>  [58] "AL-DM1(L)"         "AL-DM3(L)"         "AL-DM6(L)"        
#>  [61] "AL-D(R)"           "AL-VA1v(R)"        "AL-VA5(R)"        
#>  [64] "aL(R)"             "EB"                "NO3(L)"           
#>  [67] "aL(L)"             "INP"               "AL-DM2(R)"        
#>  [70] "AL-VM1(R)"         "POC"               "SPS(L)"           
#>  [73] "SNP(R)"            "CA(R)"             "CAN(R)"           
#>  [76] "b2(R)"             "CRE(R)"            "AL-DA3(L)"        
#>  [79] "PB(R7)"            "AL-DA2(L)"         "ICL(L)"           
#>  [82] "EPA(L)"            "AL-DA1(R)"         "AL-DL2v(R)"       
#>  [85] "AL-DL4(R)"         "a2(R)"             "ICL(R)"           
#>  [88] "PB(L4)"            "AL-VL2p(R)"        "FBl1"             
#>  [91] "AB(L)"             "SMP(R)"            "EBr3am"           
#>  [94] "AL-DM1(R)"         "LAL(R)"            "AL(R)"            
#>  [97] "FB-column3"        "ME(R)"             "AL-DC4(R)"        
#> [100] "g5(R)"             "AL-DA2(R)"         "bL(L)"            
#> [103] "AL-VL1(R)"         "hemibrain"         "FB"               
#> [106] "SCL(R)"            "gL(L)"             "EPA(R)"           
#> [109] "AL-DM3(R)"         "CRE(-ROB,-RUB)(R)" "PB(R6)"           
#> [112] "EBr2r4"            "PB(R1)"            "PB(L9)"           
#> [115] "EBr1"              "PB(L8)"            "bL(R)"            
#> [118] "AL-DP1m(R)"        "mALT(R)"           "AL-DC1(L)"        
#> [121] "FBl4"              "AL-DL1(R)"         "AL-VA7l(R)"       
#> [124] "AL-VC3m(R)"        "AL-VP4(R)"         "EBr5"             
#> [127] "ROB(R)"            "a'1(R)"            "NO2(R)"           
#> [130] "AL-VA7m(R)"        "NO3(R)"            "PB"               
#> [133] "VMNP"              "PB(R2)"            "ATL(R)"           
#> [136] "OL(R)"             "CA(L)"             "b'L(R)"           
#> [139] "SAD(-AMMC)"        "SMP(L)"            "GC"               
#> [142] "AL-DC4(L)"         "AL-DL5(L)"         "LO(R)"            
#> [145] "GOR(R)"            "AL-VC5(R)"         "AL-VM5v(R)"       
#> [148] "b'2(R)"            "PLP(R)"            "SNP(L)"           
#> [151] "AL-VP5(R)"         "AL-VP2(R)"         "BU(L)"            
#> [154] "AL-DC1(R)"         "SAD"               "AL-VM3(R)"        
#> [157] "SCL(L)"            "AL-DP1m(L)"        "AMMC"             
#> [160] "AVLP(R)"           "LX(L)"             "AL-VC4(R)"        
#> [163] "g4(R)"             "AOTU(R)"           "FBl9"             
#> [166] "AL-DA4l(R)"        "AL-DL5(R)"         "AL-DM6(R)"        
#> [169] "IB"                "LAL(L)"            "AL-DP1l(R)"       
#> [172] "AL-V(R)"           "ATL(L)"            "AL-DL2d(R)"       
#> [175] "PED(R)"            "CRE(-RUB)(L)"      "NO1(R)"           
#> [178] "PB(R4)"            "AL-VP3(R)"         "IB112"            
#> [181] "AL-DL3(R)"         "VES(R)"            "AL-VM5d(R)"       
#> [184] "VLNP(R)"           "mALT(L)"           "CX"               
#> [187] "GA(R)"             "AL-DM5(L)"         "a3(R)"            
#> [190] "AL-DM5(R)"         "LX(R)"             "AL-VA4(R)"        
#> [193] "AL-D(L)"           "AL-VM7v(R)"        "AL-VP1d(R)"       
#> [196] "NO1(L)"            "FBl7"              "MB(L)"            
#> [199] "g2(R)"             "AL-VA6(L)"         "AL-VM7d(R)"       
#> [202] "lACA(R)"           "AL-VC3l(R)"        "BU(R)"            
#> [205] "PVLP(R)"           "EBr3pw"            "AME(R)"           
#> [208] "a'L(L)"            "FBl8"              "AL-DA3(R)"        
#> [211] "b'1(R)"            "AL(L)"             "AL-VM2(R)"        
#> [214] "AL-VA1d(R)"        "PB(L7)"            "AL-DM4(R)"        
#> [217] "PB(R9)"            "MB(+ACA)(R)"       "LH(R)"            
#> [220] "AB(R)"             "SPS(R)"            "NO(L)"            
#> [223] "SIP(L)"            "a'L(R)"            "AL-DA4m(L)"       
#> [226] "AL-DM4(L)"         "SIP(R)"            "AL-VL2a(R)"       
#> [229] "RUB(R)"            "LOP(R)"            "PB(L1)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "ICL(R)"            "ME(R)"             "ATL(R)"           
#>   [4] "FBl6"              "ATL(L)"            "SNP(L)"           
#>   [7] "SIP(R)"            "NO1(L)"            "EBr5"             
#>  [10] "BU(L)"             "NO3(R)"            "g4(R)"            
#>  [13] "NO(R)"             "FBl5"              "IPS(R)"           
#>  [16] "RUB(R)"            "dACA(R)"           "AVLP(R)"          
#>  [19] "GOR(R)"            "PENP"              "a2(R)"            
#>  [22] "g2(R)"             "FBl3"              "PB(R3)"           
#>  [25] "PB(R4)"            "AL(R)"             "b'1(R)"           
#>  [28] "EBr3d"             "EBr2r4"            "SAD(-AMMC)"       
#>  [31] "b1(R)"             "AME(R)"            "AMMC"             
#>  [34] "mALT(R)"           "NO"                "FBl1"             
#>  [37] "SCL(R)"            "LAL(R)"            "a'L(L)"           
#>  [40] "PB(L3)"            "LAL(-GA)(R)"       "GC"               
#>  [43] "PB(R9)"            "a'1(R)"            "AOTU(R)"          
#>  [46] "hemibrain"         "LX(R)"             "IB"               
#>  [49] "PB(L2)"            "g1(R)"             "EBr1"             
#>  [52] "PB(L8)"            "MB(L)"             "BU(R)"            
#>  [55] "FBl2"              "NO(L)"             "SPS(R)"           
#>  [58] "PED(R)"            "WED(R)"            "PB(R1)"           
#>  [61] "b'2(R)"            "b'L(R)"            "PB(L6)"           
#>  [64] "SLP(R)"            "PB(L1)"            "PLP(R)"           
#>  [67] "a'L(R)"            "INP"               "CRE(-ROB,-RUB)(R)"
#>  [70] "EB"                "AL(L)"             "CRE(L)"           
#>  [73] "b'L(L)"            "SAD"               "SNP(R)"           
#>  [76] "LH(R)"             "FBl8"              "OL(R)"            
#>  [79] "SMP(R)"            "PB"                "g5(R)"            
#>  [82] "FBl4"              "ICL(L)"            "VES(R)"           
#>  [85] "PVLP(R)"           "NO2(R)"            "CX"               
#>  [88] "g3(R)"             "aL(R)"             "CRE(R)"           
#>  [91] "FBl7"              "PB(L5)"            "GOR(L)"           
#>  [94] "SPS(L)"            "EPA(R)"            "CAN(R)"           
#>  [97] "bL(R)"             "NO2(L)"            "SMP(L)"           
#> [100] "FBl9"              "LX(L)"             "vACA(R)"          
#> [103] "GF(R)"             "bL(L)"             "AB(R)"            
#> [106] "EPA(L)"            "RUB(L)"            "GA(R)"            
#> [109] "FB"                "ROB(R)"            "LO(R)"            
#> [112] "PB(L7)"            "a3(R)"             "EBr3am"           
#> [115] "lACA(R)"           "VMNP"              "a'2(R)"           
#> [118] "EBr6"              "NO1(R)"            "PB(R8)"           
#> [121] "VES(L)"            "LAL(L)"            "PRW"              
#> [124] "b2(R)"             "VLNP(R)"           "FLA(R)"           
#> [127] "gL(R)"             "PB(R2)"            "a1(R)"            
#> [130] "PB(L9)"            "GNG"               "PB(R5)"           
#> [133] "MB(R)"             "AB(L)"             "CA(L)"            
#> [136] "PB(R6)"            "MB(+ACA)(R)"       "CA(R)"            
#> [139] "EBr3pw"            "aL(L)"             "PB(L4)"           
#> [142] "POC"               "AOT(R)"            "SIP(L)"           
#> [145] "PB(R7)"            "NO3(L)"            "a'3(R)"           
#> [148] "mALT(L)"           "gL(L)"             "SCL(L)"           
#> [151] "LOP(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "GC"                "a'L(R)"            "AL-DL5(L)"        
#>   [4] "NO3(L)"            "AME(R)"            "VES(L)"           
#>   [7] "gL(L)"             "AL-VP5(R)"         "PB(R5)"           
#>  [10] "vACA(R)"           "NO(L)"             "AL-DL2d(R)"       
#>  [13] "AB(L)"             "OL(R)"             "EPA(R)"           
#>  [16] "FBl9"              "a'L(L)"            "EPA(L)"           
#>  [19] "AL-VA1v(R)"        "LAL(-GA)(R)"       "AL-DC1(L)"        
#>  [22] "AL-DC4(L)"         "AL-DM4(L)"         "AL-VA2(R)"        
#>  [25] "AL-DL3(R)"         "LOP(R)"            "AL-VA7m(R)"       
#>  [28] "g1(R)"             "AL-DP1m(L)"        "AL-VA7l(R)"       
#>  [31] "AL-DC2(L)"         "VLNP(R)"           "b'1(R)"           
#>  [34] "PB(L7)"            "VES(R)"            "ICL(R)"           
#>  [37] "AL-VL1(R)"         "AOTU(R)"           "NO"               
#>  [40] "EBr5"              "EBr2r4"            "g5(R)"            
#>  [43] "AL(R)"             "AL-DM5(R)"         "RUB(R)"           
#>  [46] "PB(R8)"            "FBl4"              "AL-VM3(R)"        
#>  [49] "mALT(R)"           "CRE(R)"            "ME(R)"            
#>  [52] "g4(R)"             "AL-VA6(L)"         "PB(L6)"           
#>  [55] "a'2(R)"            "AL-DM2(L)"         "FBl6"             
#>  [58] "AL-DC3(R)"         "SNP(R)"            "PB(L9)"           
#>  [61] "AL-DP1l(R)"        "AL-VA6(R)"         "AL-VM7d(R)"       
#>  [64] "AL-V(R)"           "AL-VA3(R)"         "AL-VA4(R)"        
#>  [67] "PB(L3)"            "LX(R)"             "PB(R9)"           
#>  [70] "FBl8"              "aL(L)"             "AL-DM3(R)"        
#>  [73] "AL-DM5(L)"         "AL-VA1d(R)"        "CRE(-RUB)(L)"     
#>  [76] "AL-VA5(R)"         "AL-VP1d(R)"        "POC"              
#>  [79] "BU(R)"             "AL-DA4l(R)"        "AL-DA4m(R)"       
#>  [82] "AL-DM1(R)"         "SPS(R)"            "CA(R)"            
#>  [85] "IB"                "CX"                "NO2(R)"           
#>  [88] "PB(L4)"            "SLP(R)"            "AL-VC3l(R)"       
#>  [91] "PB(L1)"            "AL-DL4(L)"         "AL-VM5d(R)"       
#>  [94] "AL-VP1m(R)"        "PED(R)"            "a'1(R)"           
#>  [97] "AB(R)"             "PB(L2)"            "PB(R1)"           
#> [100] "FBl7"              "GOR(R)"            "a3(R)"            
#> [103] "MB(L)"             "EB"                "RUB(L)"           
#> [106] "AL-VC4(R)"         "WED(R)"            "EBr6"             
#> [109] "VMNP"              "b1(R)"             "PB(L5)"           
#> [112] "PB(R3)"            "CAN(R)"            "AL-VP4(R)"        
#> [115] "a1(R)"             "PB"                "AL-VM7d(L)"       
#> [118] "LX(L)"             "AL-VL2a(R)"        "CRE(-ROB,-RUB)(R)"
#> [121] "NO1(R)"            "PLP(R)"            "NO3(R)"           
#> [124] "SPS(L)"            "g3(R)"             "FBl1"             
#> [127] "MB(R)"             "AL-DC1(R)"         "LO(R)"            
#> [130] "AL-DA3(R)"         "hemibrain"         "GOR(L)"           
#> [133] "EBr1"              "AL-DM6(L)"         "gL(R)"            
#> [136] "bL(L)"             "AL-VC1(R)"         "SMP(R)"           
#> [139] "mALT(L)"           "ATL(L)"            "b2(R)"            
#> [142] "AL(L)"             "AL-DL1(R)"         "EBr3d"            
#> [145] "AL-VM7v(R)"        "MB(+ACA)(R)"       "AL-DC2(R)"        
#> [148] "AL-VM1(R)"         "LAL(L)"            "GA(R)"            
#> [151] "lACA(R)"           "ICL(L)"            "AL-DA1(R)"        
#> [154] "FBl5"              "AL-VM5v(R)"        "b'2(R)"           
#> [157] "b'L(R)"            "AL-DA3(L)"         "SCL(L)"           
#> [160] "PENP"              "SCL(R)"            "PB(L8)"           
#> [163] "NO(R)"             "SNP(L)"            "LH(R)"            
#> [166] "PB(R4)"            "AL-VP1l(R)"        "SIP(L)"           
#> [169] "AL-VM7v(L)"        "AL-VM2(R)"         "CA(L)"            
#> [172] "dACA(R)"           "EBr3pw"            "PB(R2)"           
#> [175] "bL(R)"             "AL-DL5(R)"         "AL-VC2(R)"        
#> [178] "FB"                "g2(R)"             "AL-VC3m(R)"       
#> [181] "a2(R)"             "FBl2"              "b'L(L)"           
#> [184] "AL-DL2v(R)"        "AL-VL2p(R)"        "AL-VP3(R)"        
#> [187] "BU(L)"             "IPS(R)"            "PB(R6)"           
#> [190] "SAD(-AMMC)"        "FB-column3"        "AL-DA2(R)"        
#> [193] "AL-DC4(R)"         "NO2(L)"            "SIP(R)"           
#> [196] "AL-DM6(R)"         "SMP(L)"            "aL(R)"            
#> [199] "AL-D(L)"           "AL-DM1(L)"         "CRE(L)"           
#> [202] "PVLP(R)"           "INP"               "GNG"              
#> [205] "AL-DL4(R)"         "AL-DP1m(R)"        "AOT(R)"           
#> [208] "GF(R)"             "LAL(R)"            "EBr3am"           
#> [211] "AL-DA2(L)"         "AL-VC5(R)"         "AL-VM4(R)"        
#> [214] "FBl3"              "AL-D(R)"           "SAD"              
#> [217] "PB(R7)"            "AL-DA4m(L)"        "AL-DM4(R)"        
#> [220] "a'3(R)"            "PRW"               "ROB(R)"           
#> [223] "AL-DM2(R)"         "AMMC"              "NO1(L)"           
#> [226] "FLA(R)"            "AL-DM3(L)"         "ATL(R)"           
#> [229] "AVLP(R)"           "AL-VP2(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-VL2a(R)"        "AMMC"              "FLA(R)"           
#>   [4] "AL-DC4(L)"         "SIP(R)"            "AL-VA5(R)"        
#>   [7] "FBl6"              "AL-DM4(R)"         "AL-VC4(R)"        
#>  [10] "NO2(L)"            "a'2(R)"            "AL-VL2p(R)"       
#>  [13] "AL-VP5(R)"         "a'L(R)"            "bL(L)"            
#>  [16] "VES(R)"            "PB(R5)"            "AL-VM4(R)"        
#>  [19] "RUB(L)"            "GOR(R)"            "AOT(R)"           
#>  [22] "RUB(R)"            "EBr1"              "EPA(R)"           
#>  [25] "CA(R)"             "PB(L4)"            "PB(L7)"           
#>  [28] "CAN(R)"            "bL(R)"             "AL-DL3(R)"        
#>  [31] "AL-DA4m(R)"        "AL-VC5(R)"         "AOTU(R)"          
#>  [34] "lACA(R)"           "ME(R)"             "FBl1"             
#>  [37] "NO2(R)"            "GA(R)"             "a'1(R)"           
#>  [40] "PB(R2)"            "PB(R3)"            "AL-DL5(R)"        
#>  [43] "GF(R)"             "EBr3pw"            "MB(+ACA)(R)"      
#>  [46] "FBl9"              "mALT(R)"           "LX(R)"            
#>  [49] "AL-VA2(R)"         "AL-VP1d(R)"        "a2(R)"            
#>  [52] "SAD(-AMMC)"        "AB(R)"             "AL-DC2(R)"        
#>  [55] "AL-DL2d(R)"        "PB(L5)"            "VES(L)"           
#>  [58] "POC"               "AL-VM7v(L)"        "NO(L)"            
#>  [61] "AL-V(R)"           "PB(L1)"            "NO3(R)"           
#>  [64] "IPS(R)"            "AL-VM3(R)"         "b'1(R)"           
#>  [67] "AB(L)"             "g4(R)"             "AL-DA4l(R)"       
#>  [70] "AL-D(R)"           "AL-VA1d(R)"        "AL-VM2(R)"        
#>  [73] "SMP(L)"            "BU(L)"             "AL-DA3(L)"        
#>  [76] "AL-DA4m(L)"        "PRW"               "AL-DL2v(R)"       
#>  [79] "mALT(L)"           "a1(R)"             "AL-VC3l(R)"       
#>  [82] "AL-VM1(R)"         "aL(R)"             "NO(R)"            
#>  [85] "PB(R4)"            "AL-DC1(R)"         "AME(R)"           
#>  [88] "AL-DM3(R)"         "AL-VM7v(R)"        "CX"               
#>  [91] "PB(R6)"            "CRE(R)"            "EBr3d"            
#>  [94] "AL-DL5(L)"         "AL-VP1l(R)"        "SCL(R)"           
#>  [97] "LH(R)"             "ATL(L)"            "SNP(R)"           
#> [100] "FBl7"              "AL-DL1(R)"         "CRE(-ROB,-RUB)(R)"
#> [103] "b'L(R)"            "PB(R9)"            "AL-DC4(R)"        
#> [106] "g1(R)"             "CRE(L)"            "AL-VA7l(R)"       
#> [109] "OL(R)"             "AL-VP3(R)"         "PB(L2)"           
#> [112] "a'3(R)"            "PB(L3)"            "AL-DC2(L)"        
#> [115] "ROB(R)"            "PB(R1)"            "EBr6"             
#> [118] "PB"                "AL-DM4(L)"         "hemibrain"        
#> [121] "b1(R)"             "PB(R8)"            "gL(L)"            
#> [124] "NO3(L)"            "AL-VM5v(R)"        "AL-VA4(R)"        
#> [127] "AL(L)"             "CRE(-RUB)(L)"      "FBl3"             
#> [130] "AL-VA1v(R)"        "AL-D(L)"           "AL-VC3m(R)"       
#> [133] "EBr3am"            "b'2(R)"            "WED(R)"           
#> [136] "NO1(R)"            "AL-DM1(L)"         "AL-VA3(R)"        
#> [139] "AL-DM6(L)"         "ATL(R)"            "AVLP(R)"          
#> [142] "MB(R)"             "MB(L)"             "FBl8"             
#> [145] "AL-DL4(L)"         "AL-DA2(L)"         "AL-DP1m(L)"       
#> [148] "aL(L)"             "LAL(L)"            "AL-VA6(R)"        
#> [151] "VLNP(R)"           "CA(L)"             "gL(R)"            
#> [154] "a3(R)"             "AL-DM6(R)"         "AL-VA7m(R)"       
#> [157] "b2(R)"             "PLP(R)"            "ICL(L)"           
#> [160] "AL-VM5d(R)"        "VMNP"              "FBl4"             
#> [163] "AL(R)"             "AL-DA1(R)"         "AL-DM1(R)"        
#> [166] "LAL(-GA)(R)"       "FBl2"              "AL-DM5(L)"        
#> [169] "GNG"               "a'L(L)"            "BU(R)"            
#> [172] "PB(L9)"            "AL-DP1m(R)"        "SMP(R)"           
#> [175] "IB"                "PENP"              "AL-DM3(L)"        
#> [178] "LX(L)"             "AL-DA2(R)"         "PB(R7)"           
#> [181] "PED(R)"            "PB(L6)"            "AL-DA3(R)"        
#> [184] "EBr5"              "EB"                "AL-DM5(R)"        
#> [187] "SCL(L)"            "EPA(L)"            "g3(R)"            
#> [190] "AL-DP1l(R)"        "LAL(R)"            "AL-DC3(R)"        
#> [193] "vACA(R)"           "PB(L8)"            "AL-VM7d(L)"       
#> [196] "FB-column3"        "b'L(L)"            "AL-VC2(R)"        
#> [199] "g5(R)"             "PVLP(R)"           "FBl5"             
#> [202] "AL-DM2(R)"         "LOP(R)"            "SAD"              
#> [205] "FB"                "AL-VL1(R)"         "AL-VP4(R)"        
#> [208] "NO1(L)"            "AL-VA6(L)"         "SNP(L)"           
#> [211] "SLP(R)"            "AL-VP1m(R)"        "AL-VM7d(R)"       
#> [214] "GOR(L)"            "NO"                "LO(R)"            
#> [217] "AL-DL4(R)"         "GC"                "SPS(L)"           
#> [220] "SIP(L)"            "AL-DC1(L)"         "INP"              
#> [223] "AL-VP2(R)"         "EBr2r4"            "ICL(R)"           
#> [226] "AL-VC1(R)"         "SPS(R)"            "g2(R)"            
#> [229] "dACA(R)"           "AL-DM2(L)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method     path                               
#> [1,] "PROPFIND" "/api"                             
#> [2,] "CONNECT"  "/api/*"                           
#> [3,] "POST"     "/api/v:ver/npexplorer/rankedtable"
#> [4,] "PATCH"    "/api"                             
#> [5,] "PROPFIND" "/api/*"                           
#> [6,] "GET"      "/*"                               
#>      name                                                                
#> [1,] "github.com/labstack/echo.glob..func1"                              
#> [2,] "github.com/labstack/echo.glob..func1"                              
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [4,] "github.com/labstack/echo.glob..func1"                              
#> [5,] "github.com/labstack/echo.glob..func1"                              
#> [6,] "github.com/labstack/echo.common.static.func1"                      

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }