Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $fib19
#> $fib19$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $fib19$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $fib19$ROIs
#> [1] "LO"  "LOP" "ME" 
#> 
#> $fib19$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $fib19$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $fib19$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "PB(R6)"            "LAL(-GA)(R)"       "AL(L)"            
#>   [4] "AL-DL5(R)"         "AL-DM1(R)"         "EBr3am"           
#>   [7] "MB(L)"             "PB"                "gL(L)"            
#>  [10] "AL-DM6(R)"         "NO3(R)"            "AL-DC2(L)"        
#>  [13] "AL-DP1l(R)"        "a'3(R)"            "a'2(R)"           
#>  [16] "AL-VA3(R)"         "AL-VL1(R)"         "PRW"              
#>  [19] "PB(R1)"            "AVLP(R)"           "PB(L4)"           
#>  [22] "AL-DM5(R)"         "AL-DM1(L)"         "AL-VA7m(R)"       
#>  [25] "RUB(R)"            "PB(L5)"            "ATL(L)"           
#>  [28] "a1(R)"             "ME(R)"             "BU(R)"            
#>  [31] "CA(L)"             "PB(L6)"            "AL-DA4m(R)"       
#>  [34] "AL-DL4(R)"         "CRE(-RUB)(L)"      "a'1(R)"           
#>  [37] "MB(+ACA)(R)"       "FBl2"              "NO(L)"            
#>  [40] "AL-VM7d(R)"        "a2(R)"             "SCL(R)"           
#>  [43] "AL-VM2(R)"         "vACA(R)"           "b'L(L)"           
#>  [46] "aL(R)"             "LX(R)"             "EPA(L)"           
#>  [49] "AL-VL2p(R)"        "PED(R)"            "EBr1"             
#>  [52] "AL-DA3(L)"         "AL-VP3(R)"         "AL-VP1d(R)"       
#>  [55] "SPS(L)"            "g5(R)"             "LX(L)"            
#>  [58] "AL-VM5v(R)"        "AL-VP1l(R)"        "AL-DM5(L)"        
#>  [61] "SNP(L)"            "VES(R)"            "AL-VA6(R)"        
#>  [64] "LAL(R)"            "b2(R)"             "VMNP"             
#>  [67] "a'L(R)"            "EPA(R)"            "AL-DM4(L)"        
#>  [70] "CRE(L)"            "SMP(L)"            "a'L(L)"           
#>  [73] "SAD(-AMMC)"        "ICL(R)"            "AL-DA2(R)"        
#>  [76] "AL-DA3(R)"         "CRE(-ROB,-RUB)(R)" "AL-DA2(L)"        
#>  [79] "CA(R)"             "b1(R)"             "IB"               
#>  [82] "AL-DC2(R)"         "AL-D(L)"           "INP"              
#>  [85] "AL-VM7v(R)"        "PB(L8)"            "AL-VM1(R)"        
#>  [88] "FBl7"              "lACA(R)"           "SNP(R)"           
#>  [91] "IB112"             "FB-column3"        "RUB(L)"           
#>  [94] "a3(R)"             "AL-DA1(R)"         "AL-DA4l(R)"       
#>  [97] "AL-D(R)"           "GF(R)"             "AL-DP1m(L)"       
#> [100] "EBr3pw"            "aL(L)"             "AOTU(R)"          
#> [103] "AL-VC2(R)"         "FBl3"              "ROB(R)"           
#> [106] "AL-DL1(R)"         "AL-DL2v(R)"        "NO1(L)"           
#> [109] "FBl9"              "CAN(R)"            "AL-VA7l(R)"       
#> [112] "AL-VM5d(R)"        "SPS(R)"            "LO(R)"            
#> [115] "PB(R4)"            "AL-VA1d(R)"        "VLNP(R)"          
#> [118] "FB"                "FLA(R)"            "FBl1"             
#> [121] "AL-VC1(R)"         "AL-VM4(R)"         "SIP(L)"           
#> [124] "g2(R)"             "bL(R)"             "AL(R)"            
#> [127] "AL-DM2(R)"         "AOT(R)"            "CRE(R)"           
#> [130] "PB(R8)"            "AL-DM3(L)"         "FBl5"             
#> [133] "AL-DC1(R)"         "AL-VA5(R)"         "LH(R)"            
#> [136] "AL-VC4(R)"         "PB(L3)"            "PLP(R)"           
#> [139] "NO"                "AB(R)"             "CX"               
#> [142] "NO2(L)"            "SIP(R)"            "NO(R)"            
#> [145] "AL-V(R)"           "GA(R)"             "GC"               
#> [148] "AL-DA4m(L)"        "SLP(R)"            "dACA(R)"          
#> [151] "AL-DM2(L)"         "PB(L9)"            "GNG"              
#> [154] "g1(R)"             "PVLP(R)"           "PB(R5)"           
#> [157] "mALT(L)"           "b'2(R)"            "NO1(R)"           
#> [160] "AL-VA6(L)"         "PB(L7)"            "AB(L)"            
#> [163] "g3(R)"             "PB(R7)"            "WED(R)"           
#> [166] "FBl4"              "AL-DL5(L)"         "OL(R)"            
#> [169] "FBl6"              "PB(R2)"            "IPS(R)"           
#> [172] "AL-DM6(L)"         "AL-DC3(R)"         "SAD"              
#> [175] "AL-DC1(L)"         "gL(R)"             "hemibrain"        
#> [178] "bL(L)"             "AL-DM3(R)"         "POC"              
#> [181] "SCL(L)"            "EBr5"              "EB"               
#> [184] "FBl8"              "ATL(R)"            "PB(R3)"           
#> [187] "AL-VC5(R)"         "PB(L1)"            "AL-VM7d(L)"       
#> [190] "MB(R)"             "AL-DP1m(R)"        "AL-VC3l(R)"       
#> [193] "LOP(R)"            "NO2(R)"            "AL-VA2(R)"        
#> [196] "AL-VM3(R)"         "mALT(R)"           "ICL(L)"           
#> [199] "AL-DM4(R)"         "AL-VP5(R)"         "AL-VP1m(R)"       
#> [202] "AMMC"              "PENP"              "GOR(L)"           
#> [205] "AL-DL4(L)"         "AL-VA4(R)"         "AL-VC3m(R)"       
#> [208] "AL-VP2(R)"         "EBr2r4"            "b'1(R)"           
#> [211] "BU(L)"             "AME(R)"            "LAL(L)"           
#> [214] "b'L(R)"            "EBr3d"             "AL-DL3(R)"        
#> [217] "AL-VA1v(R)"        "AL-VL2a(R)"        "AL-DC4(L)"        
#> [220] "PB(R9)"            "NO3(L)"            "AL-DL2d(R)"       
#> [223] "AL-VP4(R)"         "SMP(R)"            "EBr6"             
#> [226] "AL-VM7v(L)"        "AL-DC4(R)"         "PB(L2)"           
#> [229] "VES(L)"            "g4(R)"             "GOR(R)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(R2)"            "VLNP(R)"           "g4(R)"            
#>   [4] "PB(L7)"            "bL(R)"             "EBr3d"            
#>   [7] "FBl2"              "CA(R)"             "a3(R)"            
#>  [10] "a'L(L)"            "IB"                "EBr5"             
#>  [13] "BU(L)"             "LOP(R)"            "gL(R)"            
#>  [16] "CRE(L)"            "FBl5"              "FBl9"             
#>  [19] "EPA(L)"            "SNP(R)"            "NO"               
#>  [22] "dACA(R)"           "SIP(R)"            "b'2(R)"           
#>  [25] "LAL(L)"            "g2(R)"             "FBl1"             
#>  [28] "SAD"               "PRW"               "ATL(R)"           
#>  [31] "IPS(R)"            "OL(R)"             "SLP(R)"           
#>  [34] "PB(R5)"            "PB(L5)"            "VMNP"             
#>  [37] "EBr2r4"            "CRE(R)"            "FLA(R)"           
#>  [40] "a'L(R)"            "VES(R)"            "LH(R)"            
#>  [43] "MB(R)"             "AOTU(R)"           "PB(R7)"           
#>  [46] "SCL(R)"            "EBr1"              "b1(R)"            
#>  [49] "SIP(L)"            "g1(R)"             "MB(L)"            
#>  [52] "b'L(R)"            "g5(R)"             "ATL(L)"           
#>  [55] "g3(R)"             "hemibrain"         "a'2(R)"           
#>  [58] "MB(+ACA)(R)"       "PB(L2)"            "WED(R)"           
#>  [61] "PLP(R)"            "LO(R)"             "POC"              
#>  [64] "PB(R1)"            "AB(R)"             "PB(L1)"           
#>  [67] "PB(L8)"            "SNP(L)"            "SAD(-AMMC)"       
#>  [70] "CRE(-ROB,-RUB)(R)" "lACA(R)"           "LAL(-GA)(R)"      
#>  [73] "a2(R)"             "PENP"              "PB(L3)"           
#>  [76] "a'3(R)"            "mALT(L)"           "PB(R8)"           
#>  [79] "b'1(R)"            "CAN(R)"            "GNG"              
#>  [82] "AL(R)"             "EBr3pw"            "FBl4"             
#>  [85] "BU(R)"             "bL(L)"             "GOR(R)"           
#>  [88] "GA(R)"             "SPS(R)"            "NO3(L)"           
#>  [91] "GC"                "gL(L)"             "LAL(R)"           
#>  [94] "PB(L4)"            "a1(R)"             "b2(R)"            
#>  [97] "ICL(R)"            "FBl3"              "AB(L)"            
#> [100] "SCL(L)"            "CA(L)"             "EBr6"             
#> [103] "PVLP(R)"           "EB"                "ROB(R)"           
#> [106] "FBl8"              "LX(L)"             "AOT(R)"           
#> [109] "RUB(R)"            "NO3(R)"            "LX(R)"            
#> [112] "ME(R)"             "ICL(L)"            "SMP(L)"           
#> [115] "NO1(R)"            "NO(L)"             "FB"               
#> [118] "AVLP(R)"           "aL(L)"             "AL(L)"            
#> [121] "NO(R)"             "EPA(R)"            "PB(R9)"           
#> [124] "PB(L9)"            "FBl6"              "NO2(R)"           
#> [127] "INP"               "PB(R6)"            "NO1(L)"           
#> [130] "VES(L)"            "PB"                "vACA(R)"          
#> [133] "CX"                "GF(R)"             "SPS(L)"           
#> [136] "PED(R)"            "GOR(L)"            "b'L(L)"           
#> [139] "EBr3am"            "PB(L6)"            "a'1(R)"           
#> [142] "RUB(L)"            "SMP(R)"            "NO2(L)"           
#> [145] "mALT(R)"           "FBl7"              "AME(R)"           
#> [148] "PB(R4)"            "AMMC"              "aL(R)"            
#> [151] "PB(R3)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "FBl3"              "FBl4"              "CA(R)"            
#>   [4] "EPA(L)"            "AL-VA2(R)"         "LAL(L)"           
#>   [7] "EBr2r4"            "AL-DA2(L)"         "g4(R)"            
#>  [10] "SAD(-AMMC)"        "CX"                "EBr6"             
#>  [13] "AMMC"              "vACA(R)"           "SAD"              
#>  [16] "AL-DA3(L)"         "FB-column3"        "SPS(L)"           
#>  [19] "AL-VP2(R)"         "bL(R)"             "a'L(L)"           
#>  [22] "PB(L5)"            "PB(L8)"            "AL(L)"            
#>  [25] "AL-VM7v(L)"        "AL-VA7m(R)"        "b2(R)"            
#>  [28] "LX(R)"             "LAL(R)"            "AL-DL5(R)"        
#>  [31] "AL-DP1l(R)"        "AL-VM7v(R)"        "CRE(R)"           
#>  [34] "INP"               "AL-DL4(R)"         "a1(R)"            
#>  [37] "NO2(L)"            "AL-DM4(L)"         "FBl8"             
#>  [40] "SCL(L)"            "GA(R)"             "MB(L)"            
#>  [43] "FBl7"              "AL-DC3(R)"         "AL-DL1(R)"        
#>  [46] "ME(R)"             "AL-DC4(L)"         "AL(R)"            
#>  [49] "AL-VC3l(R)"        "a2(R)"             "SLP(R)"           
#>  [52] "AL-DM5(R)"         "PB(R8)"            "EPA(R)"           
#>  [55] "AL-D(L)"           "GNG"               "AL-VC3m(R)"       
#>  [58] "a'3(R)"            "PB(L1)"            "PB(R4)"           
#>  [61] "PB(L4)"            "IB"                "PB(L3)"           
#>  [64] "AB(R)"             "AL-DA3(R)"         "AL-VA3(R)"        
#>  [67] "AL-DM2(L)"         "AL-VA1d(R)"        "AL-VA1v(R)"       
#>  [70] "AL-VM5d(R)"        "a'1(R)"            "FBl2"             
#>  [73] "GOR(R)"            "PB(R3)"            "AL-VP1d(R)"       
#>  [76] "RUB(R)"            "PB(R7)"            "AL-DM3(L)"        
#>  [79] "PB(L9)"            "SIP(L)"            "EBr3d"            
#>  [82] "AL-DM1(R)"         "NO3(L)"            "OL(R)"            
#>  [85] "AL-VL2a(R)"        "AL-DC2(R)"         "AL-VA5(R)"        
#>  [88] "ATL(L)"            "CA(L)"             "NO3(R)"           
#>  [91] "NO1(L)"            "MB(+ACA)(R)"       "gL(L)"            
#>  [94] "FBl1"              "AL-VA4(R)"         "b1(R)"            
#>  [97] "AL-DP1m(R)"        "PENP"              "GOR(L)"           
#> [100] "g2(R)"             "AL-VM1(R)"         "AL-VM2(R)"        
#> [103] "FBl5"              "mALT(L)"           "GF(R)"            
#> [106] "NO"                "LO(R)"             "PB(R5)"           
#> [109] "AL-DC1(L)"         "LX(L)"             "AL-VC2(R)"        
#> [112] "FBl9"              "NO2(R)"            "SNP(L)"           
#> [115] "VES(L)"            "AL-VA6(R)"         "PB(R2)"           
#> [118] "AL-DC1(R)"         "AL-DM3(R)"         "FB"               
#> [121] "AL-DL3(R)"         "AL-VA7l(R)"        "g3(R)"            
#> [124] "aL(R)"             "AL-DM4(R)"         "POC"              
#> [127] "EBr3am"            "NO(R)"             "AL-DA4m(R)"       
#> [130] "AL-VM3(R)"         "AL-DM2(R)"         "SNP(R)"           
#> [133] "EB"                "PVLP(R)"           "EBr5"             
#> [136] "AL-VC1(R)"         "AL-VC5(R)"         "AL-DM6(L)"        
#> [139] "PB(R1)"            "mALT(R)"           "AL-VC4(R)"        
#> [142] "MB(R)"             "AL-VP5(R)"         "SMP(R)"           
#> [145] "AOTU(R)"           "b'L(R)"            "g5(R)"            
#> [148] "b'1(R)"            "BU(L)"             "VES(R)"           
#> [151] "AL-VM5v(R)"        "CRE(-ROB,-RUB)(R)" "AL-DM5(L)"        
#> [154] "PB(L6)"            "ATL(R)"            "AL-DA4l(R)"       
#> [157] "a'L(R)"            "PB(R9)"            "gL(R)"            
#> [160] "AL-DL5(L)"         "AL-DA1(R)"         "b'2(R)"           
#> [163] "PRW"               "FLA(R)"            "a3(R)"            
#> [166] "AL-VM7d(R)"        "AL-VP3(R)"         "PED(R)"           
#> [169] "PB"                "VMNP"              "SMP(L)"           
#> [172] "dACA(R)"           "SCL(R)"            "AL-DM1(L)"        
#> [175] "PB(L2)"            "AL-VM4(R)"         "AL-DA4m(L)"       
#> [178] "aL(L)"             "b'L(L)"            "BU(R)"            
#> [181] "NO1(R)"            "ROB(R)"            "FBl6"             
#> [184] "AL-VP4(R)"         "GC"                "AL-DP1m(L)"       
#> [187] "AOT(R)"            "AL-DL2v(R)"        "AL-DC2(L)"        
#> [190] "ICL(L)"            "PB(L7)"            "AL-DL2d(R)"       
#> [193] "AL-D(R)"           "AL-VA6(L)"         "AME(R)"           
#> [196] "RUB(L)"            "AL-DA2(R)"         "lACA(R)"          
#> [199] "WED(R)"            "NO(L)"             "AVLP(R)"          
#> [202] "EBr1"              "a'2(R)"            "bL(L)"            
#> [205] "AL-VM7d(L)"        "ICL(R)"            "AB(L)"            
#> [208] "AL-DC4(R)"         "AL-VP1m(R)"        "VLNP(R)"          
#> [211] "g1(R)"             "PLP(R)"            "AL-DL4(L)"        
#> [214] "CRE(L)"            "LH(R)"             "PB(R6)"           
#> [217] "hemibrain"         "SPS(R)"            "CAN(R)"           
#> [220] "AL-VL1(R)"         "AL-VL2p(R)"        "LOP(R)"           
#> [223] "LAL(-GA)(R)"       "SIP(R)"            "CRE(-RUB)(L)"     
#> [226] "AL-VP1l(R)"        "AL-V(R)"           "EBr3pw"           
#> [229] "IPS(R)"            "AL-DM6(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-DM5(R)"         "AL-VP1m(R)"        "hemibrain"        
#>   [4] "EBr1"              "a'L(R)"            "AL-VC2(R)"        
#>   [7] "LAL(-GA)(R)"       "dACA(R)"           "RUB(L)"           
#>  [10] "PB(L4)"            "CRE(-ROB,-RUB)(R)" "PB(R7)"           
#>  [13] "CA(R)"             "AL-VM7v(R)"        "a'2(R)"           
#>  [16] "FBl1"              "a3(R)"             "AL-VC1(R)"        
#>  [19] "gL(R)"             "AL-DL1(R)"         "PB(R5)"           
#>  [22] "GA(R)"             "ATL(R)"            "PB(R4)"           
#>  [25] "a'L(L)"            "AL-DA4m(R)"        "GOR(R)"           
#>  [28] "AL-DP1m(R)"        "AL-VA5(R)"         "b'2(R)"           
#>  [31] "b1(R)"             "aL(L)"             "EPA(L)"           
#>  [34] "VES(L)"            "AL-DM6(R)"         "AL(R)"            
#>  [37] "AL-VM4(R)"         "AL-VA6(L)"         "IPS(R)"           
#>  [40] "AL-DA1(R)"         "AL-VP2(R)"         "ME(R)"            
#>  [43] "SNP(L)"            "LAL(R)"            "AMMC"             
#>  [46] "BU(R)"             "CRE(-RUB)(L)"      "FBl4"             
#>  [49] "FBl7"              "a2(R)"             "PRW"              
#>  [52] "AL-DA2(L)"         "FBl6"              "AL-VL1(R)"        
#>  [55] "AL-V(R)"           "SCL(L)"            "SPS(L)"           
#>  [58] "AL-DC2(R)"         "AL-VA2(R)"         "AL-VP1l(R)"       
#>  [61] "RUB(R)"            "PB(L9)"            "AL-DM1(R)"        
#>  [64] "PENP"              "NO1(L)"            "PB(L8)"           
#>  [67] "AL-VM7d(L)"        "SMP(R)"            "AL-DL4(L)"        
#>  [70] "FBl5"              "AL-DC3(R)"         "AL-DP1l(R)"       
#>  [73] "LOP(R)"            "SNP(R)"            "EBr2r4"           
#>  [76] "CRE(L)"            "PB(L3)"            "PVLP(R)"          
#>  [79] "FB-column3"        "FBl3"              "AL-DM6(L)"        
#>  [82] "ICL(L)"            "CAN(R)"            "AL-DC1(R)"        
#>  [85] "NO2(L)"            "AOTU(R)"           "g2(R)"            
#>  [88] "AL-DL5(L)"         "FBl8"              "ICL(R)"           
#>  [91] "GNG"               "g1(R)"             "NO2(R)"           
#>  [94] "MB(L)"             "AL-DA3(L)"         "AL-DA4m(L)"       
#>  [97] "AME(R)"            "AL-VA4(R)"         "a'3(R)"           
#> [100] "b'L(R)"            "AL-VC5(R)"         "AL-D(L)"          
#> [103] "AL-VA7l(R)"        "AL-VM7d(R)"        "AL-VP4(R)"        
#> [106] "PB(R6)"            "SAD"               "AL-DC1(L)"        
#> [109] "AL-VC3m(R)"        "GF(R)"             "FLA(R)"           
#> [112] "PB(R8)"            "PB"                "a'1(R)"           
#> [115] "EBr5"              "BU(L)"             "g4(R)"            
#> [118] "NO(R)"             "AL-VM5v(R)"        "b'1(R)"           
#> [121] "AL(L)"             "AL-D(R)"           "AL-VA3(R)"        
#> [124] "CA(L)"             "PB(R2)"            "OL(R)"            
#> [127] "ATL(L)"            "SCL(R)"            "AL-VP5(R)"        
#> [130] "SIP(R)"            "AL-VM5d(R)"        "mALT(R)"          
#> [133] "CRE(R)"            "PB(R3)"            "AL-DM2(R)"        
#> [136] "WED(R)"            "aL(R)"             "AL-VA1d(R)"       
#> [139] "AOT(R)"            "PLP(R)"            "IB"               
#> [142] "AL-DL2v(R)"        "AL-VM3(R)"         "PB(L1)"           
#> [145] "MB(+ACA)(R)"       "SLP(R)"            "AL-VM2(R)"        
#> [148] "SIP(L)"            "AL-DM4(L)"         "PB(L7)"           
#> [151] "FBl9"              "bL(R)"             "AL-DA2(R)"        
#> [154] "POC"               "mALT(L)"           "SMP(L)"           
#> [157] "VMNP"              "g5(R)"             "AB(R)"            
#> [160] "MB(R)"             "b2(R)"             "EPA(R)"           
#> [163] "PB(R9)"            "AL-VC4(R)"         "PB(L2)"           
#> [166] "SPS(R)"            "AL-DM1(L)"         "AL-DM5(L)"        
#> [169] "EBr3pw"            "AL-DC4(R)"         "AL-VM1(R)"        
#> [172] "AL-VP3(R)"         "PED(R)"            "GC"               
#> [175] "NO1(R)"            "SAD(-AMMC)"        "AL-DL3(R)"        
#> [178] "PB(R1)"            "AVLP(R)"           "AL-VL2a(R)"       
#> [181] "g3(R)"             "AL-DM2(L)"         "VES(R)"           
#> [184] "AL-DL2d(R)"        "AL-VP1d(R)"        "PB(L6)"           
#> [187] "LO(R)"             "LH(R)"             "b'L(L)"           
#> [190] "AL-DM4(R)"         "LX(R)"             "AL-DA4l(R)"       
#> [193] "AL-DL5(R)"         "AL-DM3(R)"         "AL-DC2(L)"        
#> [196] "NO"                "ROB(R)"            "AL-VA1v(R)"       
#> [199] "AL-VC3l(R)"        "PB(L5)"            "LAL(L)"           
#> [202] "EB"                "gL(L)"             "AL-VM7v(L)"       
#> [205] "EBr3am"            "AL-DC4(L)"         "INP"              
#> [208] "AL-DL4(R)"         "AL-VA6(R)"         "lACA(R)"          
#> [211] "AL-DM3(L)"         "NO3(L)"            "vACA(R)"          
#> [214] "AL-DA3(R)"         "GOR(L)"            "CX"               
#> [217] "EBr3d"             "FBl2"              "EBr6"             
#> [220] "AL-DP1m(L)"        "bL(L)"             "AL-VA7m(R)"       
#> [223] "AL-VL2p(R)"        "FB"                "NO3(R)"           
#> [226] "LX(L)"             "NO(L)"             "VLNP(R)"          
#> [229] "AB(L)"             "a1(R)"            
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method   path                                
#> [1,] "REPORT" "/api/*"                            
#> [2,] "GET"    "/api/v:ver/dbmeta/version"         
#> [3,] "GET"    "/api/custom/custom"                
#> [4,] "POST"   "/api/v:ver/custom/custom"          
#> [5,] "GET"    "/api/npexplorer/commonconnectivity"
#> [6,] "POST"   "/api/npexplorer/distribution"      
#>      name                                                                                              
#> [1,] "github.com/labstack/echo/v4.glob..func1"                                                         
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"                   
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCommonConnectivity-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"      
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }