Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "FLA(R)"            "ATL(R)"            "a'3(R)"           
#>   [4] "PB(R7)"            "WED(R)"            "EBr3d"            
#>   [7] "AL-DC3(R)"         "AL-VA5(R)"         "AL-VM5d(R)"       
#>  [10] "AL-DA2(L)"         "PB(L9)"            "FBl2"             
#>  [13] "FBl8"              "a3(R)"             "AL-DL3(R)"        
#>  [16] "PB(R3)"            "b1(R)"             "LX(L)"            
#>  [19] "g2(R)"             "CRE(R)"            "AL-VA7l(R)"       
#>  [22] "NO1(R)"            "PVLP(R)"           "FBl1"             
#>  [25] "SMP(R)"            "SCL(L)"            "vACA(R)"          
#>  [28] "CRE(-ROB,-RUB)(R)" "GOR(L)"            "AL-DL5(L)"        
#>  [31] "SLP(R)"            "AL-DC4(R)"         "AL-DP1m(R)"       
#>  [34] "AL-VM4(R)"         "GA(R)"             "PED(R)"           
#>  [37] "NO3(R)"            "a'2(R)"            "PB(R1)"           
#>  [40] "ICL(L)"            "FBl9"              "AL-VM5v(R)"       
#>  [43] "POC"               "LAL(-GA)(R)"       "PB"               
#>  [46] "BU(R)"             "AL-D(L)"           "IPS(R)"           
#>  [49] "AL-VC2(R)"         "AL-VP1d(R)"        "PB(L6)"           
#>  [52] "SMP(L)"            "PB(L3)"            "SAD"              
#>  [55] "AL-DA3(L)"         "AL-DC1(L)"         "AL-DM2(L)"        
#>  [58] "AL-VA6(L)"         "IB112"             "AL-DM2(R)"        
#>  [61] "EBr3pw"            "PB(L7)"            "ICL(R)"           
#>  [64] "AL-VP4(R)"         "dACA(R)"           "SAD(-AMMC)"       
#>  [67] "AL-DP1l(R)"        "AOTU(R)"           "ROB(R)"           
#>  [70] "AL-DL2v(R)"        "AL-DM3(R)"         "PRW"              
#>  [73] "AB(R)"             "PB(L4)"            "AL-DL5(R)"        
#>  [76] "LAL(L)"            "EPA(R)"            "AL-DM1(R)"        
#>  [79] "AL-VA2(R)"         "IB"                "g3(R)"            
#>  [82] "mALT(R)"           "AL-DA2(R)"         "AL-VA6(R)"        
#>  [85] "PB(L2)"            "AL-VC4(R)"         "EBr3am"           
#>  [88] "AL-DM1(L)"         "AL-VP5(R)"         "AL-DA4m(L)"       
#>  [91] "AL-VM7d(R)"        "CA(L)"             "g1(R)"            
#>  [94] "SIP(R)"            "AL-VL2a(R)"        "mALT(L)"          
#>  [97] "lACA(R)"           "AL-VL1(R)"         "AL-VM7v(R)"       
#> [100] "AOT(R)"            "PB(L1)"            "FBl4"             
#> [103] "b'2(R)"            "AL-DC2(L)"         "MB(+ACA)(R)"      
#> [106] "AL-VP3(R)"         "FB"                "a'L(R)"           
#> [109] "AL-DL4(L)"         "AL-VC3l(R)"        "LOP(R)"           
#> [112] "NO3(L)"            "PB(R4)"            "PB(R8)"           
#> [115] "NO"                "NO2(R)"            "AL-DC1(R)"        
#> [118] "AL-DL1(R)"         "AL-VM3(R)"         "EB"               
#> [121] "FBl7"              "AL-VA3(R)"         "PB(R5)"           
#> [124] "AL-DL4(R)"         "AL(L)"             "AL-VM2(R)"        
#> [127] "PB(L5)"            "PENP"              "b2(R)"            
#> [130] "FBl3"              "CX"                "aL(R)"            
#> [133] "MB(L)"             "AL-DL2d(R)"        "GF(R)"            
#> [136] "GC"                "AVLP(R)"           "AL(R)"            
#> [139] "SNP(R)"            "AL-VC5(R)"         "gL(L)"            
#> [142] "AME(R)"            "MB(R)"             "AL-VA7m(R)"       
#> [145] "AMMC"              "VLNP(R)"           "SPS(R)"           
#> [148] "EBr5"              "EBr2r4"            "AL-DP1m(L)"       
#> [151] "a'1(R)"            "bL(L)"             "AL-DC2(R)"        
#> [154] "NO2(L)"            "AL-DM6(L)"         "CRE(L)"           
#> [157] "bL(R)"             "AL-DA1(R)"         "SIP(L)"           
#> [160] "LX(R)"             "AL-VC3m(R)"        "b'1(R)"           
#> [163] "g4(R)"             "AL-VA4(R)"         "ATL(L)"           
#> [166] "VMNP"              "a1(R)"             "EBr1"             
#> [169] "PB(L8)"            "PLP(R)"            "NO(R)"            
#> [172] "EPA(L)"            "AL-V(R)"           "AL-VP1l(R)"       
#> [175] "BU(L)"             "AL-DM3(L)"         "CA(R)"            
#> [178] "LH(R)"             "aL(L)"             "AB(L)"            
#> [181] "RUB(R)"            "hemibrain"         "SPS(L)"           
#> [184] "NO(L)"             "g5(R)"             "LAL(R)"           
#> [187] "AL-VM1(R)"         "PB(R9)"            "gL(R)"            
#> [190] "CAN(R)"            "AL-VM7v(L)"        "LO(R)"            
#> [193] "b'L(R)"            "FBl5"              "AL-DM4(L)"        
#> [196] "AL-VM7d(L)"        "FB-column3"        "OL(R)"            
#> [199] "AL-DM6(R)"         "AL-VL2p(R)"        "ME(R)"            
#> [202] "AL-DC4(L)"         "INP"               "AL-DA3(R)"        
#> [205] "AL-DA4l(R)"        "AL-DA4m(R)"        "AL-DM5(R)"        
#> [208] "GNG"               "AL-D(R)"           "AL-VA1v(R)"       
#> [211] "NO1(L)"            "GOR(R)"            "AL-VP1m(R)"       
#> [214] "AL-DM5(L)"         "PB(R2)"            "VES(R)"           
#> [217] "AL-VC1(R)"         "FBl6"              "VES(L)"           
#> [220] "AL-DM4(R)"         "PB(R6)"            "SCL(R)"           
#> [223] "SNP(L)"            "AL-VA1d(R)"        "a2(R)"            
#> [226] "a'L(L)"            "CRE(-RUB)(L)"      "AL-VP2(R)"        
#> [229] "EBr6"              "RUB(L)"            "b'L(L)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "LAL(-GA)(R)"       "PB(R8)"            "NO(R)"            
#>   [4] "PB(R3)"            "lACA(R)"           "FBl3"             
#>   [7] "FBl9"              "PB(R4)"            "SLP(R)"           
#>  [10] "vACA(R)"           "NO2(L)"            "SNP(R)"           
#>  [13] "CRE(L)"            "IB"                "PB(L2)"           
#>  [16] "PB(L5)"            "g3(R)"             "AOTU(R)"          
#>  [19] "b'L(R)"            "EPA(R)"            "ATL(R)"           
#>  [22] "LX(L)"             "SPS(L)"            "FBl7"             
#>  [25] "LH(R)"             "PB(R2)"            "GA(R)"            
#>  [28] "NO"                "PB(L6)"            "EPA(L)"           
#>  [31] "LOP(R)"            "hemibrain"         "SIP(L)"           
#>  [34] "g5(R)"             "FBl2"              "VES(L)"           
#>  [37] "POC"               "a'1(R)"            "EBr3d"            
#>  [40] "a3(R)"             "PENP"              "PB(R6)"           
#>  [43] "SMP(L)"            "GOR(L)"            "a'L(R)"           
#>  [46] "g4(R)"             "AL(L)"             "SNP(L)"           
#>  [49] "FBl8"              "bL(R)"             "NO3(R)"           
#>  [52] "dACA(R)"           "FBl5"              "SAD"              
#>  [55] "EBr1"              "FBl1"              "NO2(R)"           
#>  [58] "a'3(R)"            "SPS(R)"            "g1(R)"            
#>  [61] "NO1(R)"            "AB(L)"             "SCL(L)"           
#>  [64] "CA(L)"             "b'2(R)"            "a2(R)"            
#>  [67] "g2(R)"             "EBr5"              "AVLP(R)"          
#>  [70] "OL(R)"             "LAL(L)"            "VES(R)"           
#>  [73] "INP"               "PED(R)"            "PB(L8)"           
#>  [76] "BU(R)"             "PB"                "VMNP"             
#>  [79] "EBr2r4"            "gL(L)"             "FBl4"             
#>  [82] "ICL(R)"            "aL(L)"             "NO(L)"            
#>  [85] "b'1(R)"            "LX(R)"             "PB(R1)"           
#>  [88] "PB(L7)"            "GNG"               "b'L(L)"           
#>  [91] "LAL(R)"            "PB(R5)"            "ATL(L)"           
#>  [94] "GF(R)"             "PB(R7)"            "bL(L)"            
#>  [97] "FB"                "FBl6"              "IPS(R)"           
#> [100] "CAN(R)"            "mALT(R)"           "WED(R)"           
#> [103] "gL(R)"             "EBr3pw"            "AME(R)"           
#> [106] "MB(R)"             "CRE(-ROB,-RUB)(R)" "AMMC"             
#> [109] "CX"                "ME(R)"             "PB(L4)"           
#> [112] "AL(R)"             "ROB(R)"            "a'L(L)"           
#> [115] "FLA(R)"            "PB(R9)"            "PB(L9)"           
#> [118] "SAD(-AMMC)"        "MB(+ACA)(R)"       "VLNP(R)"          
#> [121] "mALT(L)"           "PB(L3)"            "NO1(L)"           
#> [124] "b2(R)"             "a'2(R)"            "AOT(R)"           
#> [127] "GC"                "PB(L1)"            "NO3(L)"           
#> [130] "AB(R)"             "CRE(R)"            "GOR(R)"           
#> [133] "a1(R)"             "PRW"               "MB(L)"            
#> [136] "BU(L)"             "EB"                "SIP(R)"           
#> [139] "ICL(L)"            "PVLP(R)"           "b1(R)"            
#> [142] "EBr3am"            "aL(R)"             "RUB(R)"           
#> [145] "SCL(R)"            "EBr6"              "PLP(R)"           
#> [148] "LO(R)"             "CA(R)"             "RUB(L)"           
#> [151] "SMP(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-VC3l(R)"        "ICL(L)"            "FB"               
#>   [4] "VMNP"              "AL-DC4(L)"         "CRE(-ROB,-RUB)(R)"
#>   [7] "PB(R7)"            "SLP(R)"            "EBr2r4"           
#>  [10] "AL-VM7v(L)"        "AL-DL3(R)"         "AL-VA7m(R)"       
#>  [13] "AOTU(R)"           "NO(R)"             "AL-DP1m(R)"       
#>  [16] "GF(R)"             "AL-D(R)"           "PB(L7)"           
#>  [19] "NO3(L)"            "CRE(R)"            "AL(L)"            
#>  [22] "AL-VC4(R)"         "a'L(R)"            "AL-DM4(R)"        
#>  [25] "AB(L)"             "AL-VM7d(L)"        "gL(R)"            
#>  [28] "AL-VL1(R)"         "SCL(L)"            "NO1(R)"           
#>  [31] "gL(L)"             "AL-DC2(L)"         "a'3(R)"           
#>  [34] "a'2(R)"            "LH(R)"             "AL-DL4(R)"        
#>  [37] "AL-VM5d(R)"        "PB(L6)"            "AL-V(R)"          
#>  [40] "AL-DL1(R)"         "AL-VA3(R)"         "AL-VP1d(R)"       
#>  [43] "PB(R5)"            "AL-DC1(L)"         "FBl6"             
#>  [46] "a3(R)"             "AL-VA7l(R)"        "RUB(R)"           
#>  [49] "hemibrain"         "NO3(R)"            "aL(R)"            
#>  [52] "AL-VL2p(R)"        "PB(L4)"            "PRW"              
#>  [55] "AL-DA4l(R)"        "AL-VM7d(R)"        "AL-VP5(R)"        
#>  [58] "a2(R)"             "LAL(-GA)(R)"       "AL-DC3(R)"        
#>  [61] "EBr3am"            "GOR(L)"            "SAD(-AMMC)"       
#>  [64] "FBl8"              "AL(R)"             "LX(R)"            
#>  [67] "AL-DC1(R)"         "AL-VM2(R)"         "AL-DM3(L)"        
#>  [70] "AL-VP4(R)"         "AL-DM6(L)"         "SPS(L)"           
#>  [73] "b1(R)"             "AL-VP2(R)"         "ATL(L)"           
#>  [76] "GA(R)"             "AL-DM5(L)"         "AL-DA3(R)"        
#>  [79] "AL-DL2d(R)"        "AL-VC5(R)"         "AL-VM3(R)"        
#>  [82] "AL-VM4(R)"         "PB(L3)"            "PB(L8)"           
#>  [85] "PVLP(R)"           "PB(L5)"            "GC"               
#>  [88] "AL-VA6(L)"         "CRE(L)"            "VES(R)"           
#>  [91] "AB(R)"             "lACA(R)"           "AL-VC3m(R)"       
#>  [94] "EBr3pw"            "b2(R)"             "AL-DL5(L)"        
#>  [97] "PB(L9)"            "g1(R)"             "ICL(R)"           
#> [100] "AL-DA2(R)"         "PB(L2)"            "FBl1"             
#> [103] "EB"                "AL-DM4(L)"         "PB(R2)"           
#> [106] "AVLP(R)"           "LX(L)"             "NO(L)"            
#> [109] "SIP(L)"            "AL-VP3(R)"         "a1(R)"            
#> [112] "AL-VL2a(R)"        "WED(R)"            "AMMC"             
#> [115] "RUB(L)"            "AL-VM7v(R)"        "SNP(R)"           
#> [118] "g4(R)"             "FBl7"              "CAN(R)"           
#> [121] "AL-VA2(R)"         "PED(R)"            "IB"               
#> [124] "AL-VM5v(R)"        "NO"                "AL-DC2(R)"        
#> [127] "AL-DM1(R)"         "AL-DM3(R)"         "LOP(R)"           
#> [130] "EBr5"              "AL-DA2(L)"         "VES(L)"           
#> [133] "MB(R)"             "AL-DA1(R)"         "SCL(R)"           
#> [136] "AL-D(L)"           "PB(R1)"            "MB(+ACA)(R)"      
#> [139] "INP"               "AL-DM6(R)"         "b'L(R)"           
#> [142] "a'1(R)"            "FBl5"              "NO2(L)"           
#> [145] "BU(R)"             "ME(R)"             "AL-DM5(R)"        
#> [148] "AL-VP1m(R)"        "b'1(R)"            "mALT(L)"          
#> [151] "MB(L)"             "VLNP(R)"           "SNP(L)"           
#> [154] "GOR(R)"            "AL-VA1d(R)"        "FBl3"             
#> [157] "bL(L)"             "AL-VA4(R)"         "CRE(-RUB)(L)"     
#> [160] "FLA(R)"            "AL-DM1(L)"         "AL-DM2(R)"        
#> [163] "EBr6"              "AL-DL4(L)"         "FBl2"             
#> [166] "FBl4"              "AL-DA3(L)"         "PB(R6)"           
#> [169] "PB(R8)"            "PB(R9)"            "FB-column3"       
#> [172] "IPS(R)"            "AL-DA4m(R)"        "AL-VM1(R)"        
#> [175] "PENP"              "PB(R3)"            "vACA(R)"          
#> [178] "CA(L)"             "LAL(L)"            "BU(L)"            
#> [181] "CA(R)"             "AL-DL5(R)"         "POC"              
#> [184] "SMP(L)"            "AL-DL2v(R)"        "CX"               
#> [187] "AL-DA4m(L)"        "AME(R)"            "PLP(R)"           
#> [190] "AL-DM2(L)"         "FBl9"              "aL(L)"            
#> [193] "SAD"               "SIP(R)"            "g5(R)"            
#> [196] "AL-DC4(R)"         "AL-VC2(R)"         "g2(R)"            
#> [199] "ATL(R)"            "NO2(R)"            "OL(R)"            
#> [202] "AL-VA5(R)"         "AL-VC1(R)"         "LO(R)"            
#> [205] "bL(R)"             "AL-VP1l(R)"        "EPA(R)"           
#> [208] "AL-VA1v(R)"        "g3(R)"             "GNG"              
#> [211] "NO1(L)"            "a'L(L)"            "EBr1"             
#> [214] "PB(R4)"            "b'L(L)"            "AOT(R)"           
#> [217] "SPS(R)"            "PB"                "EBr3d"            
#> [220] "LAL(R)"            "dACA(R)"           "PB(L1)"           
#> [223] "EPA(L)"            "AL-DP1l(R)"        "AL-VA6(R)"        
#> [226] "SMP(R)"            "b'2(R)"            "AL-DP1m(L)"       
#> [229] "ROB(R)"            "mALT(R)"          
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-DM3(L)"         "ROB(R)"            "AL-DA4m(R)"       
#>   [4] "AL-DM1(R)"         "AL-D(L)"           "INP"              
#>   [7] "AL-VP1m(R)"        "SCL(L)"            "MB(+ACA)(R)"      
#>  [10] "FBl7"              "CRE(R)"            "b'L(R)"           
#>  [13] "AL-DM2(L)"         "AL-VA5(R)"         "PRW"              
#>  [16] "AB(L)"             "PB(L2)"            "a2(R)"            
#>  [19] "PLP(R)"            "gL(R)"             "NO(L)"            
#>  [22] "NO"                "ICL(L)"            "PB(R7)"           
#>  [25] "MB(L)"             "PB(R9)"            "AL-VA1d(R)"       
#>  [28] "OL(R)"             "AL-VC5(R)"         "NO1(R)"           
#>  [31] "PB(R1)"            "AL-DP1m(R)"        "PED(R)"           
#>  [34] "BU(L)"             "VES(R)"            "AL-VA7m(R)"       
#>  [37] "AL-VM2(R)"         "AL-DA3(L)"         "AL-DM2(R)"        
#>  [40] "AL-DP1l(R)"        "AL-VM7v(R)"        "PB(L3)"           
#>  [43] "dACA(R)"           "SCL(R)"            "g5(R)"            
#>  [46] "a3(R)"             "AL-VC1(R)"         "AL-DM6(L)"        
#>  [49] "PB(L7)"            "AL-DA1(R)"         "AL-DL2d(R)"       
#>  [52] "AL-D(R)"           "vACA(R)"           "GC"               
#>  [55] "AL-VA4(R)"         "AL-VC3l(R)"        "SMP(R)"           
#>  [58] "CRE(-ROB,-RUB)(R)" "FB"                "LX(R)"            
#>  [61] "GOR(R)"            "a'L(L)"            "PB(R6)"           
#>  [64] "g2(R)"             "PB(L1)"            "AL-DC4(R)"        
#>  [67] "AL-VL2p(R)"        "CX"                "AL-DA4m(L)"       
#>  [70] "AL-DL3(R)"         "AL-DL4(R)"         "aL(R)"            
#>  [73] "SPS(L)"            "AL-DP1m(L)"        "AL-VA1v(R)"       
#>  [76] "hemibrain"         "EBr3d"             "PB(L6)"           
#>  [79] "SNP(L)"            "FBl6"              "b1(R)"            
#>  [82] "POC"               "AL-DM1(L)"         "AL(R)"            
#>  [85] "AL-VM5d(R)"        "AL-V(R)"           "gL(L)"            
#>  [88] "FBl8"              "GOR(L)"            "FBl4"             
#>  [91] "AL-VP1l(R)"        "GF(R)"             "FBl3"             
#>  [94] "AL-VP4(R)"         "PB(R8)"            "AB(R)"            
#>  [97] "FBl1"              "LH(R)"             "ICL(R)"           
#> [100] "GNG"               "AL-DA2(R)"         "AL-DM5(R)"        
#> [103] "AL-VL1(R)"         "SIP(L)"            "a'L(R)"           
#> [106] "EPA(R)"            "AME(R)"            "aL(L)"            
#> [109] "EPA(L)"            "AL-DL1(R)"         "mALT(L)"          
#> [112] "b'2(R)"            "GA(R)"             "g1(R)"            
#> [115] "VMNP"              "VLNP(R)"           "AMMC"             
#> [118] "SIP(R)"            "CRE(L)"            "SAD(-AMMC)"       
#> [121] "FBl2"              "ATL(R)"            "g3(R)"            
#> [124] "AL-DM4(L)"         "FBl5"              "CAN(R)"           
#> [127] "AL-DA4l(R)"        "AL-VA6(L)"         "AL-DL2v(R)"       
#> [130] "AL-DL5(R)"         "RUB(L)"            "AL-VP5(R)"        
#> [133] "AL-DC2(L)"         "NO(R)"             "AVLP(R)"          
#> [136] "AL-VL2a(R)"        "AL-VM7d(R)"        "EBr3am"           
#> [139] "a'2(R)"            "AL-DL4(L)"         "IPS(R)"           
#> [142] "AL-DM4(R)"         "AOT(R)"            "IB"               
#> [145] "AL-DM5(L)"         "AL-VA2(R)"         "AL-VM3(R)"        
#> [148] "AL-VM5v(R)"        "AL-VP2(R)"         "AL-VP1d(R)"       
#> [151] "AL(L)"             "AL-VM7v(L)"        "bL(L)"            
#> [154] "AL-VA3(R)"         "PB(L5)"            "PENP"             
#> [157] "g4(R)"             "LAL(R)"            "SLP(R)"           
#> [160] "a1(R)"             "NO2(R)"            "EBr6"             
#> [163] "SNP(R)"            "BU(R)"             "a'3(R)"           
#> [166] "bL(R)"             "AL-VC4(R)"         "CA(R)"            
#> [169] "AL-DC3(R)"         "AL-VC2(R)"         "WED(R)"           
#> [172] "EBr2r4"            "mALT(R)"           "a'1(R)"           
#> [175] "b'L(L)"            "AL-DC1(R)"         "RUB(R)"           
#> [178] "CA(L)"             "NO1(L)"            "EBr5"             
#> [181] "AL-DL5(L)"         "AL-DM6(R)"         "FLA(R)"           
#> [184] "EB"                "CRE(-RUB)(L)"      "EBr3pw"           
#> [187] "lACA(R)"           "AL-VA7l(R)"        "AL-VP3(R)"        
#> [190] "LAL(-GA)(R)"       "b2(R)"             "FB-column3"       
#> [193] "FBl9"              "PB(L4)"            "AL-VC3m(R)"       
#> [196] "PB(L8)"            "AL-DC4(L)"         "PB(R3)"           
#> [199] "LOP(R)"            "LAL(L)"            "PB"               
#> [202] "PB(L9)"            "SPS(R)"            "AL-DC1(L)"        
#> [205] "LX(L)"             "SAD"               "b'1(R)"           
#> [208] "NO2(L)"            "VES(L)"            "AL-DA3(R)"        
#> [211] "AL-DC2(R)"         "PB(R5)"            "ME(R)"            
#> [214] "MB(R)"             "SMP(L)"            "AOTU(R)"          
#> [217] "AL-VM7d(L)"        "PB(R2)"            "AL-VA6(R)"        
#> [220] "AL-VM1(R)"         "ATL(L)"            "NO3(R)"           
#> [223] "LO(R)"             "PB(R4)"            "AL-DM3(R)"        
#> [226] "AL-DA2(L)"         "NO3(L)"            "EBr1"             
#> [229] "AL-VM4(R)"         "PVLP(R)"          
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                        
#> [1,] "GET"  "/api/v:ver/cached/roicompleteness"         
#> [2,] "POST" "/api/v:ver/npexplorer/roiconnectivity"     
#> [3,] "GET"  "/api/npexplorer/distribution"              
#> [4,] "GET"  "/api/serverinfo"                           
#> [5,] "GET"  "/api/v:ver/dbmeta/instances"               
#> [6,] "GET"  "/api/v:ver/skeletons/skeleton/:dataset/:id"
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }