Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DM2(L)"         "PB(L6)"            "LO(R)"            
#>   [4] "MB(+ACA)(R)"       "INP"               "NO3(R)"           
#>   [7] "vACA(R)"           "SIP(L)"            "VMNP"             
#>  [10] "AL-DC2(L)"         "AL-DM5(R)"         "AL-VP3(R)"        
#>  [13] "PB(L1)"            "SMP(L)"            "AL-VC5(R)"        
#>  [16] "PENP"              "a'1(R)"            "CA(R)"            
#>  [19] "IPS(R)"            "AL-VA1d(R)"        "AB(L)"            
#>  [22] "AL-DL5(L)"         "SPS(R)"            "LX(L)"            
#>  [25] "GNG"               "AL-DA1(R)"         "PRW"              
#>  [28] "LAL(-GA)(R)"       "ME(R)"             "EBr3am"           
#>  [31] "a'3(R)"            "g3(R)"             "AL-DA4m(L)"       
#>  [34] "VES(R)"            "FBl3"              "GOR(R)"           
#>  [37] "MB(L)"             "AL-DM1(R)"         "AL-D(R)"          
#>  [40] "AL-VA7m(R)"        "FBl7"              "AL-VA6(L)"        
#>  [43] "CRE(L)"            "AL-VP1m(R)"        "CRE(-ROB,-RUB)(R)"
#>  [46] "mALT(R)"           "IB112"             "b'1(R)"           
#>  [49] "NO2(L)"            "AL-DM4(L)"         "AL-DA4m(R)"       
#>  [52] "AL-VM1(R)"         "AL-DM5(L)"         "gL(R)"            
#>  [55] "AL-DA2(R)"         "b'2(R)"            "BU(L)"            
#>  [58] "PVLP(R)"           "AL-DL2v(R)"        "AL-VA6(R)"        
#>  [61] "CX"                "WED(R)"            "AL-VP1l(R)"       
#>  [64] "AL-DP1m(R)"        "AL-VP5(R)"         "LH(R)"            
#>  [67] "PB"                "SAD(-AMMC)"        "PB(L4)"           
#>  [70] "PB(R7)"            "FLA(R)"            "b1(R)"            
#>  [73] "MB(R)"             "AL-DM2(R)"         "AL-VM5v(R)"       
#>  [76] "GC"                "NO1(R)"            "AL-VM7d(L)"       
#>  [79] "CAN(R)"            "AL-DA4l(R)"        "AL-DP1l(R)"       
#>  [82] "AL-VC4(R)"         "VLNP(R)"           "g1(R)"            
#>  [85] "ATL(R)"            "aL(L)"             "AL-DC3(R)"        
#>  [88] "NO"                "PB(R3)"            "NO(L)"            
#>  [91] "AL-DL4(R)"         "a2(R)"             "PB(R1)"           
#>  [94] "AVLP(R)"           "GOR(L)"            "SNP(L)"           
#>  [97] "AL-DL2d(R)"        "POC"               "PB(L9)"           
#> [100] "NO1(L)"            "EPA(R)"            "CRE(-RUB)(L)"     
#> [103] "lACA(R)"           "AL-DC4(R)"         "AL-DM6(R)"        
#> [106] "dACA(R)"           "SCL(L)"            "a1(R)"            
#> [109] "AL-VL2a(R)"        "AL-VM7d(R)"        "NO(R)"            
#> [112] "AL-VA4(R)"         "FBl6"              "AL-VA7l(R)"       
#> [115] "AL-V(R)"           "AL-DM1(L)"         "AL-DA3(R)"        
#> [118] "CRE(R)"            "AL-DC2(R)"         "AL-DP1m(L)"       
#> [121] "mALT(L)"           "AL-DC4(L)"         "EBr6"             
#> [124] "NO2(R)"            "a3(R)"             "GA(R)"            
#> [127] "EBr3d"             "g4(R)"             "EPA(L)"           
#> [130] "LAL(R)"            "AL-DL3(R)"         "AL-VC3m(R)"       
#> [133] "AL-VP4(R)"         "SPS(L)"            "AL-VP1d(R)"       
#> [136] "a'2(R)"            "PED(R)"            "AL-VA1v(R)"       
#> [139] "AL-VC1(R)"         "FB"                "AL-DC1(L)"        
#> [142] "RUB(L)"            "EBr5"              "AL(R)"            
#> [145] "AL-DM3(L)"         "VES(L)"            "AL-DC1(R)"        
#> [148] "PB(L8)"            "g2(R)"             "EBr3pw"           
#> [151] "aL(R)"             "EBr1"              "ICL(L)"           
#> [154] "FBl4"              "PB(R6)"            "AL-VP2(R)"        
#> [157] "AL-VA3(R)"         "FBl8"              "b'L(L)"           
#> [160] "SLP(R)"            "AL-VA5(R)"         "PB(L5)"           
#> [163] "bL(L)"             "PB(R4)"            "AL-DL1(R)"        
#> [166] "AL-DM3(R)"         "AL-VM2(R)"         "IB"               
#> [169] "AL-DA2(L)"         "SAD"               "b'L(R)"           
#> [172] "PLP(R)"            "AL-DM6(L)"         "g5(R)"            
#> [175] "FBl2"              "bL(R)"             "GF(R)"            
#> [178] "AL-DA3(L)"         "FBl5"              "PB(R8)"           
#> [181] "SIP(R)"            "EBr2r4"            "AL-VM4(R)"        
#> [184] "SMP(R)"            "SCL(R)"            "AOTU(R)"          
#> [187] "AL-VM7v(L)"        "PB(L7)"            "AL-VL1(R)"        
#> [190] "AL-VM5d(R)"        "LAL(L)"            "AL-VM7v(R)"       
#> [193] "gL(L)"             "ROB(R)"            "FBl1"             
#> [196] "AL-DM4(R)"         "AL-VA2(R)"         "hemibrain"        
#> [199] "OL(R)"             "a'L(R)"            "SNP(R)"           
#> [202] "FB-column3"        "AL-DL4(L)"         "NO3(L)"           
#> [205] "AOT(R)"            "PB(R5)"            "AMMC"             
#> [208] "EB"                "LOP(R)"            "AL-DL5(R)"        
#> [211] "ICL(R)"            "a'L(L)"            "AL-VC2(R)"        
#> [214] "PB(L2)"            "CA(L)"             "FBl9"             
#> [217] "AL-D(L)"           "PB(R9)"            "AME(R)"           
#> [220] "AL-VC3l(R)"        "ATL(L)"            "LX(R)"            
#> [223] "RUB(R)"            "b2(R)"             "AL(L)"            
#> [226] "PB(R2)"            "AL-VM3(R)"         "PB(L3)"           
#> [229] "AB(R)"             "AL-VL2p(R)"        "BU(R)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PVLP(R)"           "CA(R)"             "EPA(L)"           
#>   [4] "LAL(L)"            "PRW"               "SIP(L)"           
#>   [7] "PED(R)"            "GC"                "GNG"              
#>  [10] "AMMC"              "PLP(R)"            "BU(R)"            
#>  [13] "PB(L4)"            "OL(R)"             "NO1(L)"           
#>  [16] "a'2(R)"            "a'1(R)"            "SAD(-AMMC)"       
#>  [19] "ATL(L)"            "PB(R8)"            "EB"               
#>  [22] "FBl7"              "LH(R)"             "CX"               
#>  [25] "mALT(R)"           "g5(R)"             "PB(R2)"           
#>  [28] "SLP(R)"            "SMP(R)"            "b'2(R)"           
#>  [31] "PB(R7)"            "MB(R)"             "AL(L)"            
#>  [34] "WED(R)"            "EBr5"              "FBl8"             
#>  [37] "SPS(R)"            "CA(L)"             "NO2(L)"           
#>  [40] "SIP(R)"            "ATL(R)"            "a'3(R)"           
#>  [43] "FBl1"              "a3(R)"             "GF(R)"            
#>  [46] "a2(R)"             "b2(R)"             "ME(R)"            
#>  [49] "g4(R)"             "GA(R)"             "LX(R)"            
#>  [52] "EBr2r4"            "LX(L)"             "PB(L8)"           
#>  [55] "LOP(R)"            "b1(R)"             "g2(R)"            
#>  [58] "SCL(R)"            "gL(R)"             "PB(L7)"           
#>  [61] "AVLP(R)"           "g3(R)"             "EBr6"             
#>  [64] "FBl4"              "CAN(R)"            "POC"              
#>  [67] "NO3(R)"            "PB(R9)"            "LO(R)"            
#>  [70] "ICL(R)"            "VLNP(R)"           "AME(R)"           
#>  [73] "RUB(R)"            "CRE(-ROB,-RUB)(R)" "EBr3am"           
#>  [76] "PB(L3)"            "CRE(R)"            "VES(R)"           
#>  [79] "NO(L)"             "PB(L2)"            "LAL(R)"           
#>  [82] "a'L(L)"            "g1(R)"             "bL(L)"            
#>  [85] "SNP(L)"            "PB(R5)"            "b'L(R)"           
#>  [88] "EPA(R)"            "PB(R1)"            "FBl9"             
#>  [91] "a1(R)"             "EBr1"              "PB(L1)"           
#>  [94] "FBl6"              "SCL(L)"            "GOR(L)"           
#>  [97] "PB(L9)"            "FBl3"              "NO2(R)"           
#> [100] "vACA(R)"           "AB(L)"             "PB(R6)"           
#> [103] "FLA(R)"            "SPS(L)"            "EBr3pw"           
#> [106] "IPS(R)"            "FBl5"              "PB(R3)"           
#> [109] "PB(L5)"            "PENP"              "aL(R)"            
#> [112] "dACA(R)"           "VMNP"              "EBr3d"            
#> [115] "SMP(L)"            "LAL(-GA)(R)"       "CRE(L)"           
#> [118] "aL(L)"             "RUB(L)"            "FB"               
#> [121] "PB"                "VES(L)"            "b'1(R)"           
#> [124] "mALT(L)"           "bL(R)"             "AOTU(R)"          
#> [127] "MB(L)"             "ROB(R)"            "NO(R)"            
#> [130] "INP"               "ICL(L)"            "FBl2"             
#> [133] "PB(R4)"            "SAD"               "BU(L)"            
#> [136] "a'L(R)"            "NO"                "PB(L6)"           
#> [139] "b'L(L)"            "AL(R)"             "GOR(R)"           
#> [142] "AOT(R)"            "IB"                "AB(R)"            
#> [145] "hemibrain"         "lACA(R)"           "NO1(R)"           
#> [148] "NO3(L)"            "MB(+ACA)(R)"       "SNP(R)"           
#> [151] "gL(L)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "ICL(R)"            "SLP(R)"            "AL-DA3(R)"        
#>   [4] "AL-DM1(L)"         "PB(L7)"            "AL-DM4(L)"        
#>   [7] "NO3(L)"            "LAL(L)"            "lACA(R)"          
#>  [10] "gL(L)"             "POC"               "SAD"              
#>  [13] "EB"                "AL-D(R)"           "AL-VC4(R)"        
#>  [16] "b2(R)"             "AL-DC3(R)"         "AL-VM7v(L)"       
#>  [19] "NO(R)"             "AL-VC1(R)"         "AL-VM4(R)"        
#>  [22] "PB(R5)"            "a'2(R)"            "MB(R)"            
#>  [25] "AL-DC2(L)"         "SAD(-AMMC)"        "VES(L)"           
#>  [28] "AL-DA4l(R)"        "CX"                "GC"               
#>  [31] "g1(R)"             "b'1(R)"            "AL-DP1m(L)"       
#>  [34] "a3(R)"             "a'3(R)"            "AME(R)"           
#>  [37] "AL-DM6(R)"         "AL-VA7m(R)"        "AL-VM7d(R)"       
#>  [40] "FBl9"              "AL-DP1l(R)"        "FBl5"             
#>  [43] "a'L(L)"            "AL-VA2(R)"         "AL-VM5v(R)"       
#>  [46] "AL-VP1l(R)"        "MB(+ACA)(R)"       "FBl4"             
#>  [49] "AL-VM3(R)"         "AB(L)"             "BU(L)"            
#>  [52] "AL-DL5(R)"         "FBl6"              "SNP(R)"           
#>  [55] "g4(R)"             "AB(R)"             "AL-DL4(R)"        
#>  [58] "AOT(R)"            "AL(R)"             "AL-DM1(R)"        
#>  [61] "AL-DL2v(R)"        "AL-VC5(R)"         "PED(R)"           
#>  [64] "FB"                "ATL(R)"            "NO2(L)"           
#>  [67] "EBr2r4"            "AL-DL5(L)"         "b'2(R)"           
#>  [70] "AL-DC4(L)"         "PB(R9)"            "LX(L)"            
#>  [73] "AL-DC2(R)"         "AL-DP1m(R)"        "CRE(-ROB,-RUB)(R)"
#>  [76] "CRE(R)"            "SCL(R)"            "EBr3d"            
#>  [79] "PVLP(R)"           "FB-column3"        "AL-DM2(R)"        
#>  [82] "PB(R7)"            "NO1(L)"            "VLNP(R)"          
#>  [85] "ROB(R)"            "AL-DL2d(R)"        "PB(R8)"           
#>  [88] "AL-VC3m(R)"        "PB(R4)"            "AL-VC3l(R)"       
#>  [91] "FLA(R)"            "PB(L9)"            "FBl2"             
#>  [94] "NO3(R)"            "g2(R)"             "RUB(L)"           
#>  [97] "EPA(R)"            "INP"               "LX(R)"            
#> [100] "NO(L)"             "AL-DL3(R)"         "AL-VP5(R)"        
#> [103] "LAL(-GA)(R)"       "VMNP"              "AL-VA7l(R)"       
#> [106] "a1(R)"             "CA(R)"             "AL-DM3(R)"        
#> [109] "AL-VL2p(R)"        "AOTU(R)"           "LO(R)"            
#> [112] "VES(R)"            "LH(R)"             "PB(L4)"           
#> [115] "AL-VP2(R)"         "SMP(L)"            "g3(R)"            
#> [118] "ME(R)"             "a'L(R)"            "AL-DL4(L)"        
#> [121] "AL-VA1d(R)"        "GA(R)"             "mALT(R)"          
#> [124] "GOR(R)"            "AL-DM5(R)"         "AL-V(R)"          
#> [127] "hemibrain"         "WED(R)"            "PB(L2)"           
#> [130] "AL-DM2(L)"         "AL-DM6(L)"         "GF(R)"            
#> [133] "AL-VP1m(R)"        "ATL(L)"            "EBr6"             
#> [136] "bL(L)"             "ICL(L)"            "AL-VM7v(R)"       
#> [139] "AL-VM7d(L)"        "NO"                "AL-DC1(R)"        
#> [142] "AL-VM5d(R)"        "mALT(L)"           "GNG"              
#> [145] "LAL(R)"            "PLP(R)"            "AL-D(L)"          
#> [148] "NO2(R)"            "AL-DA4m(R)"        "EBr5"             
#> [151] "dACA(R)"           "AL-DA3(L)"         "AL-DC4(R)"        
#> [154] "RUB(R)"            "IB"                "AL-DM5(L)"        
#> [157] "SIP(R)"            "a'1(R)"            "FBl1"             
#> [160] "FBl3"              "AL-VA6(L)"         "PB(L6)"           
#> [163] "PB(L5)"            "AL-DM3(L)"         "AL-VP4(R)"        
#> [166] "AL-VA3(R)"         "AL-VP3(R)"         "AL(L)"            
#> [169] "AL-DC1(L)"         "IPS(R)"            "EPA(L)"           
#> [172] "SCL(L)"            "CA(L)"             "aL(R)"            
#> [175] "BU(R)"             "AL-DA2(L)"         "gL(R)"            
#> [178] "CRE(-RUB)(L)"      "g5(R)"             "bL(R)"            
#> [181] "AL-VC2(R)"         "PENP"              "PRW"              
#> [184] "LOP(R)"            "SNP(L)"            "AL-DL1(R)"        
#> [187] "PB(R6)"            "NO1(R)"            "aL(L)"            
#> [190] "PB(L3)"            "SPS(L)"            "CRE(L)"           
#> [193] "EBr3pw"            "b'L(L)"            "OL(R)"            
#> [196] "SIP(L)"            "EBr1"              "AL-VA6(R)"        
#> [199] "AL-VL2a(R)"        "MB(L)"             "b1(R)"            
#> [202] "AL-VM1(R)"         "vACA(R)"           "AVLP(R)"          
#> [205] "CAN(R)"            "AL-VP1d(R)"        "PB(R3)"           
#> [208] "AL-VA1v(R)"        "FBl7"              "FBl8"             
#> [211] "AMMC"              "AL-DA4m(L)"        "PB(R2)"           
#> [214] "SPS(R)"            "PB(L1)"            "AL-DM4(R)"        
#> [217] "PB"                "AL-DA1(R)"         "AL-VA4(R)"        
#> [220] "AL-VA5(R)"         "EBr3am"            "AL-VL1(R)"        
#> [223] "SMP(R)"            "PB(R1)"            "a2(R)"            
#> [226] "PB(L8)"            "AL-DA2(R)"         "AL-VM2(R)"        
#> [229] "GOR(L)"            "b'L(R)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "PB(R7)"            "FBl2"              "AL-VL2p(R)"       
#>   [4] "PB(R6)"            "ICL(L)"            "EPA(L)"           
#>   [7] "SLP(R)"            "CX"                "LAL(-GA)(R)"      
#>  [10] "WED(R)"            "dACA(R)"           "AL-VA1d(R)"       
#>  [13] "AL-VP2(R)"         "b'L(R)"            "PB(R4)"           
#>  [16] "AL-DL2v(R)"        "AL-VM4(R)"         "SIP(L)"           
#>  [19] "SPS(L)"            "AL-VA7m(R)"        "EBr1"             
#>  [22] "gL(L)"             "GNG"               "gL(R)"            
#>  [25] "AL-DP1m(R)"        "GF(R)"             "VMNP"             
#>  [28] "PB(L6)"            "CRE(-ROB,-RUB)(R)" "a1(R)"            
#>  [31] "PB"                "AL-DM1(L)"         "INP"              
#>  [34] "AL-DA3(R)"         "AL-DL2d(R)"        "AL-DL3(R)"        
#>  [37] "AL-DL5(R)"         "AL-VC5(R)"         "AVLP(R)"          
#>  [40] "NO(L)"             "a'L(L)"            "AL-VM3(R)"        
#>  [43] "AL-VC1(R)"         "a2(R)"             "AL-DC2(L)"        
#>  [46] "AL-VA7l(R)"        "LAL(R)"            "AL-DL4(R)"        
#>  [49] "AL-VM2(R)"         "PRW"               "FBl1"             
#>  [52] "AB(R)"             "BU(L)"             "VES(L)"           
#>  [55] "AL-DC2(R)"         "b'1(R)"            "PLP(R)"           
#>  [58] "PB(R3)"            "AL-DM4(R)"         "VES(R)"           
#>  [61] "vACA(R)"           "aL(R)"             "NO1(R)"           
#>  [64] "AL-DA2(L)"         "AL-DA4m(L)"        "AL-VM7v(L)"       
#>  [67] "FBl5"              "GOR(R)"            "GOR(L)"           
#>  [70] "GC"                "PB(L9)"            "POC"              
#>  [73] "mALT(L)"           "b1(R)"             "CA(L)"            
#>  [76] "SCL(R)"            "NO"                "NO(R)"            
#>  [79] "AL(R)"             "AL-DA4l(R)"        "AL-VP1l(R)"       
#>  [82] "a'L(R)"            "BU(R)"             "LO(R)"            
#>  [85] "AL-DL4(L)"         "AL-DM3(L)"         "FBl6"             
#>  [88] "aL(L)"             "AL-DM1(R)"         "PB(R5)"           
#>  [91] "EBr5"              "PB(R8)"            "EBr3pw"           
#>  [94] "PVLP(R)"           "FBl8"              "SNP(R)"           
#>  [97] "AL-VA6(L)"         "NO3(L)"            "AL-DP1l(R)"       
#> [100] "AMMC"              "hemibrain"         "mALT(R)"          
#> [103] "PB(L1)"            "g2(R)"             "AL-DC1(L)"        
#> [106] "MB(+ACA)(R)"       "RUB(L)"            "b'2(R)"           
#> [109] "SPS(R)"            "NO3(R)"            "AL-DC3(R)"        
#> [112] "AL-VM7v(R)"        "AOTU(R)"           "EBr3d"            
#> [115] "AME(R)"            "MB(L)"             "EPA(R)"           
#> [118] "AL-DL5(L)"         "PB(L3)"            "PED(R)"           
#> [121] "CRE(R)"            "ATL(L)"            "SAD(-AMMC)"       
#> [124] "AL-VA1v(R)"        "SCL(L)"            "AL-VA3(R)"        
#> [127] "a'2(R)"            "FBl9"              "AL-DM5(R)"        
#> [130] "AOT(R)"            "PB(L5)"            "AL-DM5(L)"        
#> [133] "CRE(-RUB)(L)"      "AL-VP4(R)"         "PB(R2)"           
#> [136] "PENP"              "PB(R1)"            "AL-VA4(R)"        
#> [139] "AL-VP3(R)"         "NO1(L)"            "CA(R)"            
#> [142] "AL-DL1(R)"         "PB(R9)"            "AL-VM5d(R)"       
#> [145] "AL-VL2a(R)"        "SMP(R)"            "a'3(R)"           
#> [148] "AL-VP1m(R)"        "g1(R)"             "AL-D(L)"          
#> [151] "bL(L)"             "AL-DC4(L)"         "AL-DA2(R)"        
#> [154] "EBr3am"            "FBl3"              "IPS(R)"           
#> [157] "AL-DA3(L)"         "AL-DM2(L)"         "AL-VM1(R)"        
#> [160] "AL-VA5(R)"         "FLA(R)"            "ATL(R)"           
#> [163] "AL-VA2(R)"         "SMP(L)"            "CRE(L)"           
#> [166] "SNP(L)"            "PB(L2)"            "LOP(R)"           
#> [169] "GA(R)"             "PB(L7)"            "MB(R)"            
#> [172] "AL(L)"             "AL-VA6(R)"         "AL-VC2(R)"        
#> [175] "b2(R)"             "AL-DM3(R)"         "IB"               
#> [178] "SIP(R)"            "AL-VC3m(R)"        "AL-VM5v(R)"       
#> [181] "AL-V(R)"           "EBr6"              "PB(L8)"           
#> [184] "EB"                "AL-DM6(R)"         "AL-VL1(R)"        
#> [187] "AL-VM7d(R)"        "LX(R)"             "ME(R)"            
#> [190] "EBr2r4"            "RUB(R)"            "AB(L)"            
#> [193] "ICL(R)"            "b'L(L)"            "AL-VC4(R)"        
#> [196] "g5(R)"             "AL-DA1(R)"         "AL-DC1(R)"        
#> [199] "AL-DM2(R)"         "VLNP(R)"           "SAD"              
#> [202] "AL-DM4(L)"         "a'1(R)"            "AL-DA4m(R)"       
#> [205] "AL-VM7d(L)"        "NO2(R)"            "CAN(R)"           
#> [208] "AL-VP1d(R)"        "FBl4"              "AL-VC3l(R)"       
#> [211] "AL-VP5(R)"         "AL-DP1m(L)"        "PB(L4)"           
#> [214] "AL-DC4(R)"         "LAL(L)"            "FB-column3"       
#> [217] "bL(R)"             "OL(R)"             "g3(R)"            
#> [220] "AL-DM6(L)"         "LX(L)"             "LH(R)"            
#> [223] "a3(R)"             "g4(R)"             "ROB(R)"           
#> [226] "FBl7"              "NO2(L)"            "lACA(R)"          
#> [229] "FB"                "AL-D(R)"          
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                  
#> [1,] "POST" "/api/v:ver/npexplorer/neuronmetavals"
#> [2,] "POST" "/api/v:ver/npexplorer/rankedtable"   
#> [3,] "POST" "/api/v:ver/npexplorer/autapses"      
#> [4,] "GET"  "/api/cached/dailytype"               
#> [5,] "GET"  "/api/dbmeta/instances"               
#> [6,] "POST" "/api/npexplorer/completeness"        
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getDailyType-fm"          
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDataInstances-fm"       
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCompleteness-fm"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }