Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-VL2a(R)"        "VMNP"              "AL-VA1v(R)"       
#>   [4] "GA(R)"             "AL-DA2(R)"         "AL-DA3(R)"        
#>   [7] "AL-VA7l(R)"        "AL-DC2(L)"         "RUB(L)"           
#>  [10] "EBr6"              "MB(+ACA)(R)"       "LX(L)"            
#>  [13] "VES(L)"            "AL-VA4(R)"         "EBr3am"           
#>  [16] "SMP(L)"            "g1(R)"             "BU(R)"            
#>  [19] "CRE(-RUB)(L)"      "AL-VP1m(R)"        "a1(R)"            
#>  [22] "mALT(R)"           "SNP(R)"            "SLP(R)"           
#>  [25] "AL-VA2(R)"         "AL-VC3l(R)"        "PB(R5)"           
#>  [28] "CRE(-ROB,-RUB)(R)" "AMMC"              "SPS(R)"           
#>  [31] "VES(R)"            "NO(L)"             "MB(R)"            
#>  [34] "AL-DA4m(R)"        "AL-DM2(R)"         "a'L(L)"           
#>  [37] "b'2(R)"            "NO2(L)"            "ICL(L)"           
#>  [40] "CA(R)"             "CAN(R)"            "INP"              
#>  [43] "PB(L4)"            "PB(L3)"            "AME(R)"           
#>  [46] "mALT(L)"           "PB(L8)"            "PB(R1)"           
#>  [49] "PB(L9)"            "AL-VM3(R)"         "SAD"              
#>  [52] "FBl5"              "AL-D(R)"           "AL-VM5d(R)"       
#>  [55] "CRE(R)"            "a'L(R)"            "AL-VM7v(R)"       
#>  [58] "IPS(R)"            "AL-DA4l(R)"        "AL-DL2v(R)"       
#>  [61] "AL-VA6(R)"         "PENP"              "NO2(R)"           
#>  [64] "AL-DM5(R)"         "AL-DP1l(R)"        "GC"               
#>  [67] "LX(R)"             "PB(R9)"            "ICL(R)"           
#>  [70] "WED(R)"            "EBr5"              "AL-DC2(R)"        
#>  [73] "AL-VP5(R)"         "LO(R)"             "b1(R)"            
#>  [76] "AL-DA4m(L)"        "AL-DM6(L)"         "AL-VM7d(L)"       
#>  [79] "FBl4"              "NO(R)"             "GOR(L)"           
#>  [82] "AL(L)"             "AL-DM5(L)"         "aL(R)"            
#>  [85] "PLP(R)"            "AL-DP1m(L)"        "a'1(R)"           
#>  [88] "AL-VP3(R)"         "a2(R)"             "AL-DP1m(R)"       
#>  [91] "PB(R6)"            "g3(R)"             "LH(R)"            
#>  [94] "EPA(L)"            "AL-VC4(R)"         "SNP(L)"           
#>  [97] "SAD(-AMMC)"        "FBl7"              "AL-VM7d(R)"       
#> [100] "b2(R)"             "SPS(L)"            "PB"               
#> [103] "FB-column3"        "GNG"               "a3(R)"            
#> [106] "AL-VC3m(R)"        "AL-VM5v(R)"        "AL-DM1(L)"        
#> [109] "EBr3pw"            "FBl3"              "AL-DL2d(R)"       
#> [112] "AL-VC1(R)"         "NO3(R)"            "IB112"            
#> [115] "AL-DL4(R)"         "ATL(L)"            "NO1(R)"           
#> [118] "b'L(L)"            "AL-DL3(R)"         "CA(L)"            
#> [121] "FB"                "AL-DM2(L)"         "AL-DM1(R)"        
#> [124] "b'1(R)"            "POC"               "PB(R7)"           
#> [127] "gL(R)"             "AL-DL1(R)"         "AL-VC2(R)"        
#> [130] "AL-VL2p(R)"        "AL-VM1(R)"         "AOTU(R)"          
#> [133] "ATL(R)"            "PB(R3)"            "AVLP(R)"          
#> [136] "a'3(R)"            "CX"                "FBl2"             
#> [139] "OL(R)"             "AL-DA1(R)"         "AL-VP2(R)"        
#> [142] "PED(R)"            "AL-DA3(L)"         "PVLP(R)"          
#> [145] "AL-VM4(R)"         "AL-VP4(R)"         "b'L(R)"           
#> [148] "CRE(L)"            "FBl9"              "AL-V(R)"          
#> [151] "SIP(L)"            "EBr2r4"            "AL-VA3(R)"        
#> [154] "AL-VL1(R)"         "RUB(R)"            "LOP(R)"           
#> [157] "AL-DC4(R)"         "IB"                "PB(L2)"           
#> [160] "EPA(R)"            "AL-VA1d(R)"        "ME(R)"            
#> [163] "AL-DA2(L)"         "FBl1"              "AL-DM4(R)"        
#> [166] "PB(L5)"            "PRW"               "lACA(R)"          
#> [169] "SCL(R)"            "PB(L1)"            "PB(R8)"           
#> [172] "PB(R4)"            "AL-VA7m(R)"        "AL-VM2(R)"        
#> [175] "dACA(R)"           "EB"                "AL-DC4(L)"        
#> [178] "GOR(R)"            "AL-DM6(R)"         "AL-VA5(R)"        
#> [181] "MB(L)"             "AL-DC3(R)"         "EBr3d"            
#> [184] "AL-DM4(L)"         "ROB(R)"            "gL(L)"            
#> [187] "AB(L)"             "SCL(L)"            "FBl6"             
#> [190] "AB(R)"             "LAL(L)"            "LAL(-GA)(R)"      
#> [193] "FLA(R)"            "AL-VA6(L)"         "AL-VM7v(L)"       
#> [196] "bL(L)"             "vACA(R)"           "AL-DC1(L)"        
#> [199] "AL-DL5(L)"         "LAL(R)"            "AL-DL5(R)"        
#> [202] "VLNP(R)"           "g4(R)"             "AL-DL4(L)"        
#> [205] "AL-DM3(L)"         "PB(R2)"            "bL(R)"            
#> [208] "AL-VP1l(R)"        "hemibrain"         "g2(R)"            
#> [211] "PB(L7)"            "GF(R)"             "AL(R)"            
#> [214] "AL-VC5(R)"         "AOT(R)"            "NO1(L)"           
#> [217] "BU(L)"             "a'2(R)"            "g5(R)"            
#> [220] "FBl8"              "AL-VP1d(R)"        "EBr1"             
#> [223] "AL-D(L)"           "NO"                "NO3(L)"           
#> [226] "AL-DC1(R)"         "SIP(R)"            "PB(L6)"           
#> [229] "aL(L)"             "AL-DM3(R)"         "SMP(R)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "SCL(L)"            "AMMC"              "aL(L)"            
#>   [4] "MB(R)"             "PB(L6)"            "b'L(L)"           
#>   [7] "VES(L)"            "AOTU(R)"           "BU(L)"            
#>  [10] "a'3(R)"            "PRW"               "SIP(R)"           
#>  [13] "SNP(R)"            "FBl4"              "LH(R)"            
#>  [16] "PB(R4)"            "GNG"               "PB(R7)"           
#>  [19] "PLP(R)"            "vACA(R)"           "mALT(R)"          
#>  [22] "PB(L2)"            "FBl3"              "ATL(R)"           
#>  [25] "NO(L)"             "g1(R)"             "ICL(L)"           
#>  [28] "MB(+ACA)(R)"       "PB(R6)"            "EB"               
#>  [31] "SAD(-AMMC)"        "AB(R)"             "CA(L)"            
#>  [34] "hemibrain"         "EBr2r4"            "FBl9"             
#>  [37] "RUB(R)"            "NO1(R)"            "LX(R)"            
#>  [40] "b1(R)"             "AB(L)"             "SMP(R)"           
#>  [43] "EPA(R)"            "g5(R)"             "a'L(L)"           
#>  [46] "PB(L5)"            "EBr1"              "ME(R)"            
#>  [49] "PB(R3)"            "VLNP(R)"           "WED(R)"           
#>  [52] "VMNP"              "LO(R)"             "LOP(R)"           
#>  [55] "PB(R8)"            "SNP(L)"            "EPA(L)"           
#>  [58] "INP"               "GOR(L)"            "NO(R)"            
#>  [61] "PB(L7)"            "FBl1"              "IPS(R)"           
#>  [64] "LX(L)"             "OL(R)"             "AL(R)"            
#>  [67] "b'L(R)"            "CRE(L)"            "NO3(R)"           
#>  [70] "CRE(R)"            "PB(L9)"            "a1(R)"            
#>  [73] "LAL(L)"            "SIP(L)"            "a'L(R)"           
#>  [76] "NO"                "lACA(R)"           "SPS(L)"           
#>  [79] "g2(R)"             "PB(L1)"            "LAL(-GA)(R)"      
#>  [82] "aL(R)"             "a3(R)"             "PB(R5)"           
#>  [85] "b2(R)"             "gL(L)"             "PB(R1)"           
#>  [88] "EBr3pw"            "RUB(L)"            "EBr3am"           
#>  [91] "GF(R)"             "NO1(L)"            "CRE(-ROB,-RUB)(R)"
#>  [94] "FBl5"              "FBl8"              "AME(R)"           
#>  [97] "SMP(L)"            "SAD"               "FBl2"             
#> [100] "NO2(L)"            "PB(L3)"            "PB(R9)"           
#> [103] "g4(R)"             "gL(R)"             "a'1(R)"           
#> [106] "NO3(L)"            "VES(R)"            "ATL(L)"           
#> [109] "PB"                "EBr3d"             "AL(L)"            
#> [112] "LAL(R)"            "SLP(R)"            "POC"              
#> [115] "EBr5"              "a'2(R)"            "PB(R2)"           
#> [118] "CX"                "FB"                "bL(R)"            
#> [121] "GC"                "ROB(R)"            "FBl7"             
#> [124] "CA(R)"             "NO2(R)"            "GOR(R)"           
#> [127] "PENP"              "PED(R)"            "IB"               
#> [130] "mALT(L)"           "AOT(R)"            "b'1(R)"           
#> [133] "GA(R)"             "FLA(R)"            "ICL(R)"           
#> [136] "g3(R)"             "a2(R)"             "SCL(R)"           
#> [139] "PB(L8)"            "BU(R)"             "PVLP(R)"          
#> [142] "CAN(R)"            "b'2(R)"            "MB(L)"            
#> [145] "dACA(R)"           "bL(L)"             "PB(L4)"           
#> [148] "SPS(R)"            "EBr6"              "FBl6"             
#> [151] "AVLP(R)"          
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "ME(R)"             "gL(L)"             "PB(R1)"           
#>   [4] "AL-DM2(R)"         "GC"                "AL-VC2(R)"        
#>   [7] "AL-VM2(R)"         "CRE(-ROB,-RUB)(R)" "SMP(R)"           
#>  [10] "WED(R)"            "SPS(L)"            "AL-DP1m(L)"       
#>  [13] "SNP(L)"            "LH(R)"             "AL-DM3(R)"        
#>  [16] "AL-VA7m(R)"        "NO2(L)"            "NO3(R)"           
#>  [19] "FBl9"              "aL(L)"             "AB(L)"            
#>  [22] "a1(R)"             "LAL(L)"            "FB"               
#>  [25] "EBr1"              "FB-column3"        "INP"              
#>  [28] "g2(R)"             "AL-DM2(L)"         "AL-VA3(R)"        
#>  [31] "AL-VA7l(R)"        "AOT(R)"            "PB(R6)"           
#>  [34] "a'1(R)"            "CA(R)"             "PB(R2)"           
#>  [37] "b'2(R)"            "EBr5"              "AL-DA4m(R)"       
#>  [40] "GA(R)"             "AL-VC1(R)"         "RUB(R)"           
#>  [43] "AL-DA3(R)"         "AL-VM7v(R)"        "lACA(R)"          
#>  [46] "AL-DL5(R)"         "AL-VL2a(R)"        "CRE(R)"           
#>  [49] "AL-DM4(L)"         "AL-DL1(R)"         "AL-VM3(R)"        
#>  [52] "LOP(R)"            "LAL(R)"            "PED(R)"           
#>  [55] "EBr6"              "PB(L1)"            "PB"               
#>  [58] "BU(R)"             "NO(R)"             "LO(R)"            
#>  [61] "a'L(L)"            "AMMC"              "AL-DA3(L)"        
#>  [64] "CRE(-RUB)(L)"      "PB(R3)"            "OL(R)"            
#>  [67] "g1(R)"             "EPA(R)"            "AL-VC4(R)"        
#>  [70] "AL-DC1(R)"         "EBr3d"             "PB(L6)"           
#>  [73] "AL-VA1v(R)"        "AL-VP1m(R)"        "PB(L2)"           
#>  [76] "AL-D(L)"           "MB(+ACA)(R)"       "FBl4"             
#>  [79] "AL-DP1l(R)"        "vACA(R)"           "SMP(L)"           
#>  [82] "LX(R)"             "ROB(R)"            "PB(R8)"           
#>  [85] "AL-DA4l(R)"        "AL-DC2(R)"         "AL-DM1(R)"        
#>  [88] "AL-VM7d(L)"        "PB(L7)"            "RUB(L)"           
#>  [91] "BU(L)"             "SIP(R)"            "AL-DL2d(R)"       
#>  [94] "AL-VP4(R)"         "SAD"               "PLP(R)"           
#>  [97] "PVLP(R)"           "FBl7"              "b'L(L)"           
#> [100] "AL-VA1d(R)"        "AL-VM5v(R)"        "dACA(R)"          
#> [103] "a'L(R)"            "AL-VM1(R)"         "AOTU(R)"          
#> [106] "NO1(L)"            "b2(R)"             "EB"               
#> [109] "FBl8"              "VES(L)"            "IB"               
#> [112] "mALT(R)"           "EBr2r4"            "AL-DA1(R)"        
#> [115] "aL(R)"             "AB(R)"             "SIP(L)"           
#> [118] "AL-DA2(L)"         "SNP(R)"            "PB(L4)"           
#> [121] "AL-VA6(R)"         "g5(R)"             "NO2(R)"           
#> [124] "AL-DL3(R)"         "AL-VL1(R)"         "VMNP"             
#> [127] "AVLP(R)"           "AL-DL4(R)"         "AL-VA5(R)"        
#> [130] "EBr3am"            "AL(R)"             "AL-VM7d(R)"       
#> [133] "PB(R4)"            "AL-VC3l(R)"        "PB(L3)"           
#> [136] "EBr3pw"            "FBl2"              "AL-D(R)"          
#> [139] "AL-DM3(L)"         "ICL(L)"            "AL-DC4(R)"        
#> [142] "LAL(-GA)(R)"       "SLP(R)"            "AL-DM6(R)"        
#> [145] "AL-VM4(R)"         "FBl1"              "AL-DA2(R)"        
#> [148] "IPS(R)"            "EPA(L)"            "AL-VA4(R)"        
#> [151] "AL-VC5(R)"         "SPS(R)"            "b'L(R)"           
#> [154] "PB(L8)"            "g4(R)"             "GNG"              
#> [157] "FBl5"              "LX(L)"             "AL-VM5d(R)"       
#> [160] "b1(R)"             "GF(R)"             "hemibrain"        
#> [163] "AL-DC2(L)"         "AL-DL4(L)"         "PB(L9)"           
#> [166] "CRE(L)"            "ICL(R)"            "b'1(R)"           
#> [169] "GOR(L)"            "bL(R)"             "MB(R)"            
#> [172] "AL-DL2v(R)"        "NO1(R)"            "NO"               
#> [175] "AME(R)"            "PB(L5)"            "a2(R)"            
#> [178] "ATL(R)"            "ATL(L)"            "PB(R9)"           
#> [181] "VES(R)"            "AL-VP3(R)"         "a'3(R)"           
#> [184] "FBl6"              "AL-VA6(L)"         "AL-V(R)"          
#> [187] "AL-DC3(R)"         "AL-DM5(R)"         "a3(R)"            
#> [190] "VLNP(R)"           "PB(R5)"            "a'2(R)"           
#> [193] "AL-DC4(L)"         "AL-DM6(L)"         "CAN(R)"           
#> [196] "AL-VP1d(R)"        "POC"               "AL-VM7v(L)"       
#> [199] "gL(R)"             "bL(L)"             "SCL(R)"           
#> [202] "SAD(-AMMC)"        "GOR(R)"            "NO(L)"            
#> [205] "AL-VP2(R)"         "PRW"               "mALT(L)"          
#> [208] "CA(L)"             "g3(R)"             "MB(L)"            
#> [211] "AL-VP1l(R)"        "AL(L)"             "AL-DA4m(L)"       
#> [214] "AL-DM1(L)"         "FBl3"              "SCL(L)"           
#> [217] "CX"                "AL-VA2(R)"         "AL-VL2p(R)"       
#> [220] "AL-DM5(L)"         "NO3(L)"            "PENP"             
#> [223] "AL-DP1m(R)"        "AL-VP5(R)"         "FLA(R)"           
#> [226] "AL-DM4(R)"         "AL-VC3m(R)"        "PB(R7)"           
#> [229] "AL-DC1(L)"         "AL-DL5(L)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "LAL(-GA)(R)"       "dACA(R)"           "AL-DA3(L)"        
#>   [4] "gL(R)"             "AL-VA7m(R)"        "CA(R)"            
#>   [7] "GOR(R)"            "NO1(L)"            "AL-DA4m(L)"       
#>  [10] "AL-DC2(L)"         "PB(R4)"            "AL-VA1d(R)"       
#>  [13] "PB(L1)"            "a'2(R)"            "PB(L4)"           
#>  [16] "AL-DA4l(R)"        "GOR(L)"            "g2(R)"            
#>  [19] "PLP(R)"            "FBl6"              "CAN(R)"           
#>  [22] "AB(L)"             "NO3(R)"            "PB(R1)"           
#>  [25] "EPA(L)"            "AOTU(R)"           "AVLP(R)"          
#>  [28] "AL-DC3(R)"         "PB(R6)"            "NO3(L)"           
#>  [31] "FBl8"              "FBl9"              "AL-VM7v(R)"       
#>  [34] "PB(R5)"            "PB(L9)"            "AL-VM2(R)"        
#>  [37] "GA(R)"             "LX(R)"             "EPA(R)"           
#>  [40] "LH(R)"             "AL-DL5(L)"         "SIP(R)"           
#>  [43] "PB(R3)"            "AL-VP1m(R)"        "FBl3"             
#>  [46] "NO2(R)"            "AL-DA4m(R)"        "AL-DP1m(R)"       
#>  [49] "AL-DM3(R)"         "AL-DL4(L)"         "FBl2"             
#>  [52] "b1(R)"             "LAL(R)"            "AL-VA4(R)"        
#>  [55] "aL(R)"             "MB(L)"             "BU(L)"            
#>  [58] "AL-DM3(L)"         "FBl4"              "AL-DM4(R)"        
#>  [61] "AL-DM5(R)"         "AL-DM6(R)"         "AL-VM7v(L)"       
#>  [64] "AL-VL1(R)"         "VLNP(R)"           "PED(R)"           
#>  [67] "GC"                "PB(L8)"            "PB(R9)"           
#>  [70] "ATL(L)"            "a'1(R)"            "AL-VM5v(R)"       
#>  [73] "CRE(R)"            "PB(R8)"            "AL-VA6(L)"        
#>  [76] "INP"               "NO(L)"             "AL-DA1(R)"        
#>  [79] "AL-DM2(R)"         "AL-VM3(R)"         "AL-VP3(R)"        
#>  [82] "PB(L5)"            "AL-DA2(L)"         "AL-DC4(L)"        
#>  [85] "ATL(R)"            "mALT(R)"           "a'L(R)"           
#>  [88] "AL-DM1(L)"         "AL-DM5(L)"         "SNP(L)"           
#>  [91] "AL-VA6(R)"         "AL-VM1(R)"         "AL-VM5d(R)"       
#>  [94] "POC"               "SAD"               "VMNP"             
#>  [97] "AL-DM6(L)"         "AL-DA3(R)"         "a'3(R)"           
#> [100] "NO2(L)"            "EBr5"              "bL(R)"            
#> [103] "AL-D(L)"           "AL-DC2(R)"         "AL-V(R)"          
#> [106] "CRE(-ROB,-RUB)(R)" "g4(R)"             "SMP(R)"           
#> [109] "b'1(R)"            "NO1(R)"            "SMP(L)"           
#> [112] "hemibrain"         "b2(R)"             "EBr2r4"           
#> [115] "FBl1"              "a3(R)"             "IB"               
#> [118] "PRW"               "AL-VP1l(R)"        "g5(R)"            
#> [121] "SLP(R)"            "AL-VC3l(R)"        "AL-VM4(R)"        
#> [124] "AL-VM7d(R)"        "WED(R)"            "AL-VL2a(R)"       
#> [127] "FLA(R)"            "AL-D(R)"           "LOP(R)"           
#> [130] "NO"                "LO(R)"             "MB(+ACA)(R)"      
#> [133] "FB-column3"        "ME(R)"             "a'L(L)"           
#> [136] "AL-VP4(R)"         "FB"                "BU(R)"            
#> [139] "PB(L3)"            "EB"                "CX"               
#> [142] "EBr1"              "SPS(L)"            "AL-DP1m(L)"       
#> [145] "NO(R)"             "AL-VC2(R)"         "IPS(R)"           
#> [148] "AL-DL2d(R)"        "AL-VA7l(R)"        "AL-VC5(R)"        
#> [151] "AL-VP1d(R)"        "b'L(R)"            "PB(L6)"           
#> [154] "SAD(-AMMC)"        "PVLP(R)"           "AL-DL1(R)"        
#> [157] "mALT(L)"           "gL(L)"             "MB(R)"            
#> [160] "AL-DL4(R)"         "AL-DM1(R)"         "PB(L2)"           
#> [163] "FBl5"              "AME(R)"            "SNP(R)"           
#> [166] "ICL(L)"            "LX(L)"             "SCL(R)"           
#> [169] "EBr3pw"            "ICL(R)"            "AL-DL5(R)"        
#> [172] "g1(R)"             "aL(L)"             "b'L(L)"           
#> [175] "AL-DC1(R)"         "GF(R)"             "AL-DM4(L)"        
#> [178] "EBr3am"            "CA(L)"             "SIP(L)"           
#> [181] "ROB(R)"            "bL(L)"             "RUB(R)"           
#> [184] "AL-DM2(L)"         "CRE(-RUB)(L)"      "AL-VA3(R)"        
#> [187] "PB"                "AB(R)"             "AL-DL3(R)"        
#> [190] "AL-VC3m(R)"        "AL-VM7d(L)"        "AL-VA5(R)"        
#> [193] "PB(R7)"            "VES(R)"            "PB(R2)"           
#> [196] "AL-VC4(R)"         "EBr6"              "EBr3d"            
#> [199] "AL-DC1(L)"         "GNG"               "AOT(R)"           
#> [202] "AMMC"              "LAL(L)"            "AL(L)"            
#> [205] "g3(R)"             "CRE(L)"            "FBl7"             
#> [208] "AL(R)"             "AL-DL2v(R)"        "AL-VA1v(R)"       
#> [211] "AL-VP5(R)"         "SPS(R)"            "lACA(R)"          
#> [214] "VES(L)"            "AL-VP2(R)"         "vACA(R)"          
#> [217] "a1(R)"             "PENP"              "AL-DC4(R)"        
#> [220] "b'2(R)"            "a2(R)"             "PB(L7)"           
#> [223] "OL(R)"             "AL-DA2(R)"         "RUB(L)"           
#> [226] "AL-VL2p(R)"        "AL-DP1l(R)"        "AL-VA2(R)"        
#> [229] "AL-VC1(R)"         "SCL(L)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method  path                          
#> [1,] "TRACE" "/api/*"                      
#> [2,] "GET"   "/api/npexplorer/neuronmeta"  
#> [3,] "POST"  "/api/npexplorer/distribution"
#> [4,] "POST"  "/api/raw/cypher/transaction" 
#> [5,] "GET"   "/api/cached/roicompleteness" 
#> [6,] "GET"   "/api"                        
#>      name                                                                                        
#> [1,] "github.com/labstack/echo/v4.glob..func1"                                                   
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMeta-fm"  
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROICompleteness-fm"    
#> [6,] "github.com/labstack/echo/v4.glob..func1"                                                   
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }