Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "LH(R)"             "PB(R2)"            "bL(R)"            
#>   [4] "AL-VP3(R)"         "AL-VP1d(R)"        "CRE(L)"           
#>   [7] "AL-DA1(R)"         "AL-VC3l(R)"        "RUB(R)"           
#>  [10] "RUB(L)"            "IB112"             "AL-DC1(R)"        
#>  [13] "AL-VC3m(R)"        "AL-VL1(R)"         "AL-VM7d(L)"       
#>  [16] "AL-DA2(L)"         "NO"                "NO2(L)"           
#>  [19] "AOTU(R)"           "PB(L6)"            "FBl5"             
#>  [22] "SLP(R)"            "AL-DC1(L)"         "AL-DM6(L)"        
#>  [25] "PB(R1)"            "b'L(L)"            "AL-DC3(R)"        
#>  [28] "GC"                "FBl3"              "a3(R)"            
#>  [31] "AL-D(L)"           "CX"                "PVLP(R)"          
#>  [34] "EBr2r4"            "AL-VA2(R)"         "AL-VA3(R)"        
#>  [37] "AL-VA6(R)"         "NO(R)"             "AL-DM3(L)"        
#>  [40] "ATL(R)"            "CAN(R)"            "PB(R4)"           
#>  [43] "AL-DA4m(R)"        "CA(L)"             "b'2(R)"           
#>  [46] "AL(L)"             "AL-DL4(L)"         "AL-DM4(L)"        
#>  [49] "SIP(R)"            "gL(R)"             "OL(R)"            
#>  [52] "AL-DL2d(R)"        "EBr1"              "AL-DL2v(R)"       
#>  [55] "MB(R)"             "AL-DC2(R)"         "AL-VA1v(R)"       
#>  [58] "b'1(R)"            "PB(L4)"            "PB"               
#>  [61] "AL-DA2(R)"         "AL-DP1m(R)"        "AL-VA5(R)"        
#>  [64] "AL-VP4(R)"         "NO1(R)"            "b2(R)"            
#>  [67] "a'L(R)"            "AL-DL4(R)"         "AL-DM2(R)"        
#>  [70] "AL-DP1l(R)"        "PB(R7)"            "SAD(-AMMC)"       
#>  [73] "AL-VL2a(R)"        "VLNP(R)"           "ROB(R)"           
#>  [76] "a'1(R)"            "g5(R)"             "FBl1"             
#>  [79] "AVLP(R)"           "ICL(R)"            "g2(R)"            
#>  [82] "AL-DA3(L)"         "EBr3d"             "AL-D(R)"          
#>  [85] "VMNP"              "PB(R8)"            "AL-VA6(L)"        
#>  [88] "FBl7"              "b1(R)"             "LAL(R)"           
#>  [91] "PENP"              "PB(L9)"            "AL-V(R)"          
#>  [94] "a'2(R)"            "INP"               "SNP(R)"           
#>  [97] "AME(R)"            "AL-DA4l(R)"        "AL-VM3(R)"        
#> [100] "dACA(R)"           "AL-DL1(R)"         "WED(R)"           
#> [103] "NO(L)"             "a'L(L)"            "BU(R)"            
#> [106] "EPA(R)"            "ME(R)"             "b'L(R)"           
#> [109] "BU(L)"             "SPS(L)"            "AL-DL5(R)"        
#> [112] "SMP(L)"            "IB"                "AL-VM2(R)"        
#> [115] "NO3(R)"            "LX(R)"             "bL(L)"            
#> [118] "GOR(R)"            "AL-VM7v(R)"        "SIP(L)"           
#> [121] "g4(R)"             "a2(R)"             "AL-VA4(R)"        
#> [124] "AL-VC2(R)"         "AL-VL2p(R)"        "PLP(R)"           
#> [127] "AL-VM7v(L)"        "AL-DL3(R)"         "lACA(R)"          
#> [130] "FB"                "LO(R)"             "AL-VP2(R)"        
#> [133] "mALT(L)"           "PB(L3)"            "AL-VA7m(R)"       
#> [136] "AL-DM4(R)"         "PB(L8)"            "AL(R)"            
#> [139] "AL-VP1m(R)"        "SCL(L)"            "NO1(L)"           
#> [142] "LAL(-GA)(R)"       "GF(R)"             "AL-VM5d(R)"       
#> [145] "vACA(R)"           "LAL(L)"            "AL-VC4(R)"        
#> [148] "GOR(L)"            "VES(R)"            "AB(R)"            
#> [151] "AMMC"              "NO3(L)"            "AL-DA4m(L)"       
#> [154] "AL-DL5(L)"         "SNP(L)"            "LX(L)"            
#> [157] "GNG"               "AL-DM5(R)"         "hemibrain"        
#> [160] "AL-DA3(R)"         "AL-VM5v(R)"        "AL-VM7d(R)"       
#> [163] "AL-VP1l(R)"        "SMP(R)"            "a'3(R)"           
#> [166] "NO2(R)"            "EBr5"              "AL-DM1(L)"        
#> [169] "ICL(L)"            "g1(R)"             "FBl4"             
#> [172] "AL-VP5(R)"         "AL-DC2(L)"         "FBl2"             
#> [175] "FBl6"              "AL-VC1(R)"         "FLA(R)"           
#> [178] "FBl8"              "AL-DM1(R)"         "PB(R5)"           
#> [181] "CRE(-ROB,-RUB)(R)" "CRE(R)"            "AL-DM3(R)"        
#> [184] "PB(L5)"            "MB(+ACA)(R)"       "SPS(R)"           
#> [187] "mALT(R)"           "SCL(R)"            "AL-DM5(L)"        
#> [190] "GA(R)"             "EB"                "aL(L)"            
#> [193] "PED(R)"            "PB(R9)"            "AL-DC4(R)"        
#> [196] "AL-VM4(R)"         "EBr6"              "EBr3am"           
#> [199] "PB(L1)"            "PB(R3)"            "AL-VM1(R)"        
#> [202] "LOP(R)"            "FBl9"              "EBr3pw"           
#> [205] "AOT(R)"            "AL-DM2(L)"         "PRW"              
#> [208] "gL(L)"             "AB(L)"             "g3(R)"            
#> [211] "CA(R)"             "AL-VA1d(R)"        "aL(R)"            
#> [214] "ATL(L)"            "AL-DM6(R)"         "MB(L)"            
#> [217] "CRE(-RUB)(L)"      "AL-VA7l(R)"        "PB(L2)"           
#> [220] "SAD"               "FB-column3"        "IPS(R)"           
#> [223] "POC"               "PB(R6)"            "AL-VC5(R)"        
#> [226] "AL-DP1m(L)"        "PB(L7)"            "EPA(L)"           
#> [229] "VES(L)"            "a1(R)"             "AL-DC4(L)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(R5)"            "IB"                "mALT(R)"          
#>   [4] "g2(R)"             "NO1(R)"            "FBl1"             
#>   [7] "FBl9"              "AME(R)"            "PB(L2)"           
#>  [10] "CRE(-ROB,-RUB)(R)" "PB(R9)"            "PB(L9)"           
#>  [13] "FBl2"              "AOT(R)"            "PB(L3)"           
#>  [16] "WED(R)"            "b'L(R)"            "SAD(-AMMC)"       
#>  [19] "LAL(-GA)(R)"       "LH(R)"             "PB(L4)"           
#>  [22] "LAL(R)"            "AB(L)"             "BU(L)"            
#>  [25] "VMNP"              "SNP(R)"            "EPA(R)"           
#>  [28] "SIP(R)"            "ICL(L)"            "LOP(R)"           
#>  [31] "bL(R)"             "AMMC"              "NO3(R)"           
#>  [34] "EBr5"              "a'2(R)"            "LO(R)"            
#>  [37] "NO2(R)"            "INP"               "ME(R)"            
#>  [40] "g1(R)"             "vACA(R)"           "SPS(R)"           
#>  [43] "FB"                "EBr2r4"            "g4(R)"            
#>  [46] "BU(R)"             "PLP(R)"            "EBr3d"            
#>  [49] "EPA(L)"            "EB"                "mALT(L)"          
#>  [52] "SMP(L)"            "dACA(R)"           "PB(L8)"           
#>  [55] "CAN(R)"            "AL(R)"             "PB"               
#>  [58] "PB(L7)"            "PB(R4)"            "b'L(L)"           
#>  [61] "PB(R6)"            "SAD"               "gL(R)"            
#>  [64] "SNP(L)"            "FBl4"              "VLNP(R)"          
#>  [67] "VES(L)"            "GA(R)"             "GOR(L)"           
#>  [70] "FBl8"              "AB(R)"             "GNG"              
#>  [73] "GOR(R)"            "g3(R)"             "gL(L)"            
#>  [76] "FBl5"              "FBl7"              "a1(R)"            
#>  [79] "PB(R2)"            "ICL(R)"            "LX(L)"            
#>  [82] "MB(R)"             "POC"               "RUB(R)"           
#>  [85] "LAL(L)"            "PED(R)"            "b2(R)"            
#>  [88] "CRE(L)"            "g5(R)"             "SLP(R)"           
#>  [91] "NO2(L)"            "PB(L1)"            "NO(R)"            
#>  [94] "PVLP(R)"           "ATL(R)"            "OL(R)"            
#>  [97] "b'2(R)"            "aL(R)"             "MB(L)"            
#> [100] "a'L(R)"            "bL(L)"             "FBl3"             
#> [103] "FLA(R)"            "MB(+ACA)(R)"       "PB(R3)"           
#> [106] "SCL(R)"            "a'L(L)"            "AOTU(R)"          
#> [109] "SIP(L)"            "GC"                "AL(L)"            
#> [112] "PENP"              "SPS(L)"            "PB(R8)"           
#> [115] "VES(R)"            "hemibrain"         "PB(R7)"           
#> [118] "CRE(R)"            "AVLP(R)"           "aL(L)"            
#> [121] "a'3(R)"            "a2(R)"             "PRW"              
#> [124] "NO3(L)"            "b'1(R)"            "PB(L5)"           
#> [127] "PB(L6)"            "EBr3pw"            "NO1(L)"           
#> [130] "NO"                "a'1(R)"            "RUB(L)"           
#> [133] "CX"                "NO(L)"             "ATL(L)"           
#> [136] "LX(R)"             "ROB(R)"            "PB(R1)"           
#> [139] "FBl6"              "CA(R)"             "IPS(R)"           
#> [142] "CA(L)"             "GF(R)"             "EBr1"             
#> [145] "EBr6"              "b1(R)"             "SMP(R)"           
#> [148] "EBr3am"            "lACA(R)"           "a3(R)"            
#> [151] "SCL(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "INP"               "AL-DL2d(R)"        "AL-DL4(R)"        
#>   [4] "AL-VC3l(R)"        "AL-DL5(L)"         "PB(R1)"           
#>   [7] "gL(R)"             "b'1(R)"            "PB(L5)"           
#>  [10] "mALT(L)"           "PED(R)"            "AL-VC4(R)"        
#>  [13] "PB(R2)"            "PB(L4)"            "PB"               
#>  [16] "FBl2"              "ATL(R)"            "VES(L)"           
#>  [19] "NO1(L)"            "SNP(L)"            "aL(R)"            
#>  [22] "FBl1"              "AL-VM7d(R)"        "ME(R)"            
#>  [25] "a'L(L)"            "AL-DC1(L)"         "AL-DC3(R)"        
#>  [28] "AL-VA6(R)"         "g2(R)"             "PB(R8)"           
#>  [31] "LAL(R)"            "PB(R6)"            "PB(L1)"           
#>  [34] "FBl9"              "ICL(R)"            "LAL(-GA)(R)"      
#>  [37] "AL-DP1m(L)"        "LAL(L)"            "AL-DM5(L)"        
#>  [40] "AL-DC1(R)"         "SNP(R)"            "VES(R)"           
#>  [43] "PVLP(R)"           "FBl6"              "AL-VM1(R)"        
#>  [46] "PB(L2)"            "EBr1"              "NO1(R)"           
#>  [49] "AB(L)"             "AOTU(R)"           "AL-VM3(R)"        
#>  [52] "CAN(R)"            "AOT(R)"            "EBr3am"           
#>  [55] "mALT(R)"           "EB"                "EBr3pw"           
#>  [58] "g5(R)"             "IPS(R)"            "LX(L)"            
#>  [61] "IB"                "AL-DA3(L)"         "AL-DL4(L)"        
#>  [64] "NO3(R)"            "AL-VM7d(L)"        "NO(R)"            
#>  [67] "NO2(L)"            "AL-DP1m(R)"        "AL-VL1(R)"        
#>  [70] "GF(R)"             "FB-column3"        "PB(R4)"           
#>  [73] "GNG"               "AL-DL5(R)"         "RUB(R)"           
#>  [76] "AL-VM7v(L)"        "CRE(L)"            "AL-VP3(R)"        
#>  [79] "GA(R)"             "vACA(R)"           "MB(+ACA)(R)"      
#>  [82] "AL-VA7l(R)"        "AL-V(R)"           "PB(R7)"           
#>  [85] "gL(L)"             "AL-VP1d(R)"        "WED(R)"           
#>  [88] "AL-DA1(R)"         "PLP(R)"            "PB(L6)"           
#>  [91] "bL(L)"             "PRW"               "g1(R)"            
#>  [94] "AL-DM3(R)"         "SMP(L)"            "SAD(-AMMC)"       
#>  [97] "AL-DM1(R)"         "AL-VC1(R)"         "AME(R)"           
#> [100] "AL-DA3(R)"         "FBl5"              "b1(R)"            
#> [103] "NO2(R)"            "AL-VA1d(R)"        "AL-DM1(L)"        
#> [106] "SIP(R)"            "aL(L)"             "AL-VA5(R)"        
#> [109] "CRE(-ROB,-RUB)(R)" "EBr2r4"            "NO3(L)"           
#> [112] "AL-VA7m(R)"        "AL-VC3m(R)"        "NO(L)"            
#> [115] "AL-VA2(R)"         "GOR(L)"            "LX(R)"            
#> [118] "PB(R9)"            "AL-DP1l(R)"        "GC"               
#> [121] "PB(L8)"            "FBl3"              "AL-VL2p(R)"       
#> [124] "b'L(R)"            "AL-DA4m(R)"        "AL-VP4(R)"        
#> [127] "PB(R5)"            "PB(L3)"            "SPS(R)"           
#> [130] "AL-VP1l(R)"        "a'3(R)"            "EBr6"             
#> [133] "FBl7"              "AL-DA2(R)"         "EPA(L)"           
#> [136] "RUB(L)"            "AL-VM5v(R)"        "AL-VM7v(R)"       
#> [139] "AMMC"              "SIP(L)"            "a3(R)"            
#> [142] "dACA(R)"           "AL-DM4(R)"         "g3(R)"            
#> [145] "a2(R)"             "FLA(R)"            "GOR(R)"           
#> [148] "SCL(L)"            "PENP"              "CA(R)"            
#> [151] "AVLP(R)"           "b'L(L)"            "AL-DL3(R)"        
#> [154] "AL-DA4m(L)"        "AL-VA6(L)"         "BU(R)"            
#> [157] "AL-VA1v(R)"        "g4(R)"             "FBl4"             
#> [160] "AL-DL2v(R)"        "AL-DC2(L)"         "AL-DC4(R)"        
#> [163] "AL-DA2(L)"         "AL-D(L)"           "LO(R)"            
#> [166] "AL-D(R)"           "lACA(R)"           "EBr3d"            
#> [169] "SCL(R)"            "AL-DC2(R)"         "AL-DM6(R)"        
#> [172] "AL-VC2(R)"         "AL-VL2a(R)"        "VMNP"             
#> [175] "a'1(R)"            "AB(R)"             "AL-VM2(R)"        
#> [178] "AL-VP1m(R)"        "hemibrain"         "SPS(L)"           
#> [181] "AL-VP2(R)"         "CA(L)"             "FBl8"             
#> [184] "AL-VA3(R)"         "CRE(-RUB)(L)"      "AL-DL1(R)"        
#> [187] "AL-DM5(R)"         "AL-VM4(R)"         "ATL(L)"           
#> [190] "CX"                "CRE(R)"            "BU(L)"            
#> [193] "NO"                "SLP(R)"            "a'L(R)"           
#> [196] "AL-DM6(L)"         "ICL(L)"            "b'2(R)"           
#> [199] "AL-DM4(L)"         "bL(R)"             "MB(R)"            
#> [202] "AL-DM2(R)"         "AL-VM5d(R)"        "AL-VP5(R)"        
#> [205] "a'2(R)"            "PB(R3)"            "PB(L7)"           
#> [208] "AL(R)"             "AL-VA4(R)"         "LOP(R)"           
#> [211] "a1(R)"             "SAD"               "FB"               
#> [214] "LH(R)"             "POC"               "SMP(R)"           
#> [217] "b2(R)"             "AL-DM2(L)"         "MB(L)"            
#> [220] "AL-DM3(L)"         "AL-DC4(L)"         "PB(L9)"           
#> [223] "VLNP(R)"           "EPA(R)"            "AL-VC5(R)"        
#> [226] "ROB(R)"            "OL(R)"             "AL-DA4l(R)"       
#> [229] "EBr5"              "AL(L)"            
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.2`
#> $`hemibrain:v1.2`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2`$ROIs
#>   [1] "SPS(R)"            "ME(R)"             "AL-DC2(L)"        
#>   [4] "AL-VM5d(R)"        "PB(R6)"            "FLA(R)"           
#>   [7] "gL(R)"             "CRE(-RUB)(L)"      "LAL(R)"           
#>  [10] "AL-DC2(R)"         "AL-VC4(R)"         "AL-VP5(R)"        
#>  [13] "AL-VC3l(R)"        "AL-VP2(R)"         "a'2(R)"           
#>  [16] "AOT(R)"            "b1(R)"             "bL(R)"            
#>  [19] "AL-DL1(R)"         "AL-DM2(R)"         "AL-VA3(R)"        
#>  [22] "PB(L2)"            "a1(R)"             "NO1(L)"           
#>  [25] "EPA(L)"            "b'L(L)"            "mALT(L)"          
#>  [28] "GF(R)"             "NO"                "FBl6"             
#>  [31] "SAD"               "FB"                "ATL(R)"           
#>  [34] "AL-DP1m(L)"        "AL-DA3(R)"         "dACA(R)"          
#>  [37] "PB(L8)"            "AL-DM2(L)"         "LO(R)"            
#>  [40] "AL-DL2v(R)"        "AL-VA4(R)"         "AMMC"             
#>  [43] "FBl7"              "IB"                "AL-DL5(L)"        
#>  [46] "a'1(R)"            "INP"               "AL-VL2p(R)"       
#>  [49] "g1(R)"             "VMNP"              "AL(L)"            
#>  [52] "PB(L6)"            "VES(L)"            "NO(R)"            
#>  [55] "IPS(R)"            "AL-DC4(R)"         "SMP(R)"           
#>  [58] "PRW"               "EPA(R)"            "AL-DL4(R)"        
#>  [61] "AL-VM3(R)"         "LX(R)"             "AL-DA4m(L)"       
#>  [64] "PB(L9)"            "PB(R4)"            "MB(L)"            
#>  [67] "SLP(R)"            "AL-DM4(R)"         "AL-DP1l(R)"       
#>  [70] "AL-DM4(L)"         "FBl4"              "PB(R3)"           
#>  [73] "AL-DP1m(R)"        "bL(L)"             "FBl5"             
#>  [76] "RUB(R)"            "vACA(R)"           "SPS(L)"           
#>  [79] "ROB(R)"            "GOR(R)"            "EB"               
#>  [82] "AL-VP1d(R)"        "b2(R)"             "LAL(L)"           
#>  [85] "OL(R)"             "EBr3d"             "AL-DM5(L)"        
#>  [88] "CAN(R)"            "NO2(L)"            "SIP(R)"           
#>  [91] "PB(R9)"            "AL-DC3(R)"         "CX"               
#>  [94] "SIP(L)"            "g5(R)"             "LX(L)"            
#>  [97] "AL-VA1d(R)"        "AL-VA6(R)"         "AL-DM5(R)"        
#> [100] "SCL(R)"            "EBr2r4"            "a'L(R)"           
#> [103] "g4(R)"             "FBl2"              "GNG"              
#> [106] "AL-DL5(R)"         "SCL(L)"            "b'L(R)"           
#> [109] "NO3(L)"            "MB(+ACA)(R)"       "a'L(L)"           
#> [112] "AL-DM6(R)"         "VLNP(R)"           "PB(R5)"           
#> [115] "mALT(R)"           "AL-DM3(L)"         "PB(L3)"           
#> [118] "FB-column3"        "AL-VL2a(R)"        "LAL(-GA)(R)"      
#> [121] "g2(R)"             "LH(R)"             "AL-VM7d(R)"       
#> [124] "g3(R)"             "PB(L7)"            "aL(L)"            
#> [127] "AL-DL3(R)"         "AL-VL1(R)"         "AL-VP1l(R)"       
#> [130] "a2(R)"             "PB"                "AL-VM5v(R)"       
#> [133] "GC"                "RUB(L)"            "a3(R)"            
#> [136] "NO(L)"             "AL-DA2(R)"         "AL-DC4(L)"        
#> [139] "AL-DM1(L)"         "NO2(R)"            "AL-VA1v(R)"       
#> [142] "CA(L)"             "GA(R)"             "lACA(R)"          
#> [145] "AL-VA6(L)"         "AVLP(R)"           "b'2(R)"           
#> [148] "hemibrain"         "SNP(L)"            "AL-DA4m(R)"       
#> [151] "AL-D(R)"           "AL-VP1m(R)"        "PENP"             
#> [154] "CA(R)"             "AB(R)"             "WED(R)"           
#> [157] "PB(R2)"            "PED(R)"            "AL-DC1(L)"        
#> [160] "EBr3pw"            "FBl3"              "PB(L5)"           
#> [163] "BU(L)"             "BU(R)"             "VES(R)"           
#> [166] "AL-DA1(R)"         "SNP(R)"            "AL-DL4(L)"        
#> [169] "PB(L4)"            "NO1(R)"            "PB(R8)"           
#> [172] "AL-DM6(L)"         "AL-VA2(R)"         "AL-VP3(R)"        
#> [175] "PB(L1)"            "PVLP(R)"           "AL-VA7l(R)"       
#> [178] "AL-DA3(L)"         "AL-DA4l(R)"        "AL-VM4(R)"        
#> [181] "EBr3am"            "EBr5"              "CRE(L)"           
#> [184] "FBl9"              "AL-VC1(R)"         "AL-VC3m(R)"       
#> [187] "AL-VM2(R)"         "CRE(-ROB,-RUB)(R)" "ICL(L)"           
#> [190] "FBl8"              "aL(R)"             "PLP(R)"           
#> [193] "ICL(R)"            "NO3(R)"            "AL-VM7v(L)"       
#> [196] "AME(R)"            "AL-DL2d(R)"        "AL-DM3(R)"        
#> [199] "AL-VC5(R)"         "POC"               "EBr6"             
#> [202] "LOP(R)"            "SMP(L)"            "AL-DA2(L)"        
#> [205] "AL-D(L)"           "SAD(-AMMC)"        "AL(R)"            
#> [208] "AL-DM1(R)"         "AL-V(R)"           "AL-VP4(R)"        
#> [211] "ATL(L)"            "PB(R1)"            "AL-VM7v(R)"       
#> [214] "AB(L)"             "GOR(L)"            "MB(R)"            
#> [217] "AL-VC2(R)"         "AL-VM1(R)"         "EBr1"             
#> [220] "AL-DC1(R)"         "b'1(R)"            "a'3(R)"           
#> [223] "CRE(R)"            "AL-VM7d(L)"        "FBl1"             
#> [226] "AL-VA5(R)"         "AOTU(R)"           "PB(R7)"           
#> [229] "gL(L)"             "AL-VA7m(R)"       
#> 
#> $`hemibrain:v1.2`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method     path                                        
#> [1,] "HEAD"     "/api/*"                                    
#> [2,] "PROPFIND" "/api/*"                                    
#> [3,] "GET"      "/api/npexplorer/findneurons"               
#> [4,] "GET"      "/api/v:ver/raw/keyvalue/key/:instance/:key"
#> [5,] "POST"     "/api/v:ver/npexplorer/rankedtable"         
#> [6,] "POST"     "/api/v:ver/skeletons/skeleton/:dataset/:id"
#>      name                                                                                       
#> [1,] "github.com/labstack/echo.glob..func1"                                                     
#> [2,] "github.com/labstack/echo.glob..func1"                                                     
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getFindNeurons-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }