Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-VM7v(L)"        "a'1(R)"            "AL-DM1(R)"        
#>   [4] "LOP(R)"            "AL-DC3(R)"         "EBr2r4"           
#>   [7] "AL-DC2(L)"         "FBl7"              "CA(R)"            
#>  [10] "aL(L)"             "INP"               "AL-DC1(R)"        
#>  [13] "AL-VC2(R)"         "CRE(-ROB,-RUB)(R)" "LAL(L)"           
#>  [16] "AL-VA6(R)"         "mALT(L)"           "WED(R)"           
#>  [19] "BU(L)"             "dACA(R)"           "PB(L4)"           
#>  [22] "AL-DL2d(R)"        "AL-DM6(R)"         "CA(L)"            
#>  [25] "EB"                "AL-VA6(L)"         "AL-DA2(R)"        
#>  [28] "AL-VP1m(R)"        "FBl5"              "RUB(L)"           
#>  [31] "SPS(R)"            "aL(R)"             "FLA(R)"           
#>  [34] "AL-DM6(L)"         "NO"                "CRE(-RUB)(L)"     
#>  [37] "AL-DP1m(R)"        "POC"               "NO2(L)"           
#>  [40] "AL-DM2(L)"         "bL(L)"             "VES(L)"           
#>  [43] "hemibrain"         "g1(R)"             "SNP(R)"           
#>  [46] "IPS(R)"            "AL-VP3(R)"         "PB(L5)"           
#>  [49] "a'L(L)"            "AL-VL1(R)"         "gL(R)"            
#>  [52] "FBl9"              "g3(R)"             "PB(L9)"           
#>  [55] "EBr3pw"            "PB(R4)"            "AL-DL1(R)"        
#>  [58] "AL-DC4(L)"         "AL-VA1d(R)"        "SCL(R)"           
#>  [61] "GA(R)"             "SAD(-AMMC)"        "GNG"              
#>  [64] "EPA(R)"            "AL-DM4(L)"         "EPA(L)"           
#>  [67] "AL-DM5(R)"         "AL-VA7m(R)"        "a1(R)"            
#>  [70] "SNP(L)"            "AL-DM2(R)"         "CX"               
#>  [73] "PB(R6)"            "LAL(-GA)(R)"       "EBr1"             
#>  [76] "AL-VC4(R)"         "g2(R)"             "AL-DM3(L)"        
#>  [79] "AL-VM7d(L)"        "PB(R9)"            "AL-VM7d(R)"       
#>  [82] "AL-DA3(L)"         "AVLP(R)"           "AL-DL2v(R)"       
#>  [85] "a2(R)"             "gL(L)"             "AL-DM5(L)"        
#>  [88] "AL-DP1m(L)"        "PB(L6)"            "g5(R)"            
#>  [91] "AL-DA3(R)"         "PB(L8)"            "SIP(R)"           
#>  [94] "CAN(R)"            "AL-DA1(R)"         "AL-VL2a(R)"       
#>  [97] "AL-VP1l(R)"        "AL-DL5(L)"         "AL(R)"            
#> [100] "AL-VA4(R)"         "PB(L2)"            "VMNP"             
#> [103] "AL-DA4m(R)"        "EBr6"              "ICL(L)"           
#> [106] "AL-DM4(R)"         "SPS(L)"            "b1(R)"            
#> [109] "IB"                "AL-D(L)"           "LX(L)"            
#> [112] "SMP(R)"            "ROB(R)"            "AL-DL4(R)"        
#> [115] "IB112"             "AL-VA7l(R)"        "AL-VM2(R)"        
#> [118] "FBl6"              "GOR(R)"            "AL-DC2(R)"        
#> [121] "VES(R)"            "bL(R)"             "b'L(L)"           
#> [124] "PENP"              "CRE(R)"            "NO(R)"            
#> [127] "SAD"               "AB(R)"             "LAL(R)"           
#> [130] "AL-D(R)"           "AOT(R)"            "PRW"              
#> [133] "FBl4"              "PB(R2)"            "RUB(R)"           
#> [136] "a'2(R)"            "FBl2"              "a3(R)"            
#> [139] "AL-VM4(R)"         "AMMC"              "GOR(L)"           
#> [142] "AL-DM1(L)"         "PB(L7)"            "AL-VM3(R)"        
#> [145] "AOTU(R)"           "PLP(R)"            "PB(R1)"           
#> [148] "AL-VP4(R)"         "ME(R)"             "AL-VP2(R)"        
#> [151] "lACA(R)"           "PED(R)"            "AL-DA2(L)"        
#> [154] "AL-VL2p(R)"        "GF(R)"             "a'L(R)"           
#> [157] "AL-DC1(L)"         "AME(R)"            "OL(R)"            
#> [160] "AL-VM5v(R)"        "AL-DA4m(L)"        "NO3(L)"           
#> [163] "LH(R)"             "AB(L)"             "b'2(R)"           
#> [166] "SMP(L)"            "PB(L3)"            "mALT(R)"          
#> [169] "PVLP(R)"           "NO(L)"             "AL-VA3(R)"        
#> [172] "MB(+ACA)(R)"       "AL-DC4(R)"         "b'L(R)"           
#> [175] "PB(L1)"            "AL-DA4l(R)"        "AL-VM5d(R)"       
#> [178] "AL-VM7v(R)"        "PB"                "g4(R)"            
#> [181] "AL-VC3m(R)"        "SCL(L)"            "NO2(R)"           
#> [184] "AL-DL5(R)"         "AL-DP1l(R)"        "VLNP(R)"          
#> [187] "vACA(R)"           "GC"                "NO3(R)"           
#> [190] "AL-DL4(L)"         "AL-DL3(R)"         "SIP(L)"           
#> [193] "NO1(L)"            "PB(R8)"            "LO(R)"            
#> [196] "AL-VC3l(R)"        "AL-VM1(R)"         "a'3(R)"           
#> [199] "ICL(R)"            "LX(R)"             "FBl1"             
#> [202] "FBl3"              "SLP(R)"            "ATL(L)"           
#> [205] "FBl8"              "AL-VP1d(R)"        "b'1(R)"           
#> [208] "PB(R3)"            "AL-DM3(R)"         "EBr3am"           
#> [211] "MB(L)"             "AL-VP5(R)"         "b2(R)"            
#> [214] "FB-column3"        "AL-VC1(R)"         "EBr5"             
#> [217] "EBr3d"             "AL-VA1v(R)"        "AL-V(R)"          
#> [220] "AL-VA2(R)"         "PB(R5)"            "PB(R7)"           
#> [223] "BU(R)"             "ATL(R)"            "FB"               
#> [226] "NO1(R)"            "AL(L)"             "CRE(L)"           
#> [229] "MB(R)"             "AL-VA5(R)"         "AL-VC5(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "AOTU(R)"           "EBr5"              "NO"               
#>   [4] "PB(R1)"            "a3(R)"             "RUB(R)"           
#>   [7] "a'L(R)"            "g5(R)"             "b1(R)"            
#>  [10] "AL(R)"             "a'3(R)"            "EBr1"             
#>  [13] "dACA(R)"           "VES(R)"            "OL(R)"            
#>  [16] "SLP(R)"            "SCL(L)"            "SNP(R)"           
#>  [19] "LO(R)"             "PB(R2)"            "PB(L5)"           
#>  [22] "g3(R)"             "PB(L3)"            "NO(R)"            
#>  [25] "RUB(L)"            "EBr3am"            "SAD"              
#>  [28] "EBr2r4"            "FBl3"              "ICL(R)"           
#>  [31] "NO2(L)"            "PB(R7)"            "aL(R)"            
#>  [34] "EPA(R)"            "FBl1"              "FBl2"             
#>  [37] "PB(R6)"            "LAL(-GA)(R)"       "NO3(R)"           
#>  [40] "ICL(L)"            "AB(R)"             "CAN(R)"           
#>  [43] "a'L(L)"            "PB(L1)"            "AVLP(R)"          
#>  [46] "b'2(R)"            "GF(R)"             "CA(R)"            
#>  [49] "b'L(L)"            "SPS(R)"            "GOR(L)"           
#>  [52] "NO1(L)"            "g4(R)"             "ROB(R)"           
#>  [55] "FBl6"              "PB(R3)"            "POC"              
#>  [58] "AOT(R)"            "SMP(R)"            "g1(R)"            
#>  [61] "MB(L)"             "PB(L7)"            "ATL(L)"           
#>  [64] "lACA(R)"           "PB(L8)"            "MB(+ACA)(R)"      
#>  [67] "GOR(R)"            "a2(R)"             "PRW"              
#>  [70] "CRE(R)"            "gL(R)"             "PB(L9)"           
#>  [73] "SAD(-AMMC)"        "LAL(L)"            "BU(L)"            
#>  [76] "NO3(L)"            "SIP(L)"            "MB(R)"            
#>  [79] "CX"                "g2(R)"             "BU(R)"            
#>  [82] "AL(L)"             "CRE(L)"            "ATL(R)"           
#>  [85] "GNG"               "bL(R)"             "ME(R)"            
#>  [88] "AME(R)"            "INP"               "PB(L4)"           
#>  [91] "SMP(L)"            "PED(R)"            "PB"               
#>  [94] "mALT(L)"           "PB(R4)"            "PENP"             
#>  [97] "EBr6"              "VMNP"              "EBr3d"            
#> [100] "bL(L)"             "EBr3pw"            "EPA(L)"           
#> [103] "NO(L)"             "GA(R)"             "FLA(R)"           
#> [106] "PLP(R)"            "gL(L)"             "LH(R)"            
#> [109] "PB(L2)"            "LX(R)"             "a'2(R)"           
#> [112] "EB"                "PVLP(R)"           "aL(L)"            
#> [115] "b'1(R)"            "a1(R)"             "mALT(R)"          
#> [118] "b2(R)"             "SCL(R)"            "IPS(R)"           
#> [121] "PB(R5)"            "IB"                "AMMC"             
#> [124] "SNP(L)"            "FBl5"              "LAL(R)"           
#> [127] "hemibrain"         "GC"                "FB"               
#> [130] "FBl4"              "VLNP(R)"           "CA(L)"            
#> [133] "SIP(R)"            "FBl7"              "LOP(R)"           
#> [136] "NO1(R)"            "PB(R9)"            "a'1(R)"           
#> [139] "FBl8"              "FBl9"              "LX(L)"            
#> [142] "vACA(R)"           "WED(R)"            "b'L(R)"           
#> [145] "SPS(L)"            "PB(L6)"            "NO2(R)"           
#> [148] "CRE(-ROB,-RUB)(R)" "AB(L)"             "PB(R8)"           
#> [151] "VES(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "FBl5"              "LX(L)"             "AL-VA5(R)"        
#>   [4] "SCL(L)"            "a'3(R)"            "a'L(R)"           
#>   [7] "ICL(L)"            "AB(R)"             "AL-DM4(R)"        
#>  [10] "FLA(R)"            "VMNP"              "BU(R)"            
#>  [13] "AL-V(R)"           "CRE(-ROB,-RUB)(R)" "WED(R)"           
#>  [16] "AL-DL1(R)"         "AL-VM3(R)"         "hemibrain"        
#>  [19] "PB(L8)"            "AL-DL3(R)"         "POC"              
#>  [22] "EBr3am"            "AL-DA1(R)"         "AL-DC2(R)"        
#>  [25] "SNP(L)"            "LO(R)"             "AL-VA6(R)"        
#>  [28] "GF(R)"             "ROB(R)"            "AL-VP2(R)"        
#>  [31] "AL-DL4(L)"         "gL(R)"             "bL(R)"            
#>  [34] "AL-DM1(R)"         "AL-VM7d(R)"        "PB(L5)"           
#>  [37] "FBl7"              "AL-DM3(R)"         "FBl2"             
#>  [40] "LH(R)"             "IPS(R)"            "AL-DA2(R)"        
#>  [43] "AL-DL2d(R)"        "AL-VC4(R)"         "g2(R)"            
#>  [46] "AL-VM7d(L)"        "AL-VL2a(R)"        "SIP(R)"           
#>  [49] "VES(L)"            "AL-VM4(R)"         "g1(R)"            
#>  [52] "AL-VA6(L)"         "gL(L)"             "MB(R)"            
#>  [55] "NO3(R)"            "AOT(R)"            "LOP(R)"           
#>  [58] "SPS(R)"            "AL-VC5(R)"         "PB(L4)"           
#>  [61] "AL-DP1l(R)"        "AL-VC3m(R)"        "AL-DA3(L)"        
#>  [64] "GNG"               "PB(R4)"            "g3(R)"            
#>  [67] "PB(L1)"            "CX"                "SAD"              
#>  [70] "BU(L)"             "AL-VA1v(R)"        "AL-VA4(R)"        
#>  [73] "AL-VA7l(R)"        "ATL(L)"            "FB-column3"       
#>  [76] "OL(R)"             "a3(R)"             "EBr3d"            
#>  [79] "g4(R)"             "AL-DM2(R)"         "PENP"             
#>  [82] "AL-VM7v(L)"        "VES(R)"            "AL-DA3(R)"        
#>  [85] "AL-DP1m(L)"        "PB(L7)"            "PVLP(R)"          
#>  [88] "EPA(L)"            "NO2(L)"            "SIP(L)"           
#>  [91] "EBr2r4"            "PED(R)"            "b'L(R)"           
#>  [94] "AMMC"              "AL(R)"             "AL-VC3l(R)"       
#>  [97] "GOR(R)"            "AL-DC3(R)"         "AL-VA2(R)"        
#> [100] "a1(R)"             "PB(L3)"            "PB(R3)"           
#> [103] "AL-VP4(R)"         "b'L(L)"            "NO(L)"            
#> [106] "NO2(R)"            "PB(L2)"            "RUB(R)"           
#> [109] "INP"               "AL-VL1(R)"         "AL(L)"            
#> [112] "AL-DL5(R)"         "CA(L)"             "CRE(L)"           
#> [115] "FBl9"              "AL-VM1(R)"         "a2(R)"            
#> [118] "NO"                "GC"                "SAD(-AMMC)"       
#> [121] "AL-VL2p(R)"        "AL-VM2(R)"         "AL-DM1(L)"        
#> [124] "FBl3"              "CA(R)"             "PB(R6)"           
#> [127] "AL-DC4(L)"         "a'2(R)"            "AL-DC2(L)"        
#> [130] "AL-DL5(L)"         "AL-VP1m(R)"        "CRE(R)"           
#> [133] "SLP(R)"            "SMP(L)"            "b'1(R)"           
#> [136] "CAN(R)"            "AL-DM6(L)"         "EBr3pw"           
#> [139] "MB(+ACA)(R)"       "FBl6"              "FBl8"             
#> [142] "LAL(R)"            "AL-VP3(R)"         "GOR(L)"           
#> [145] "CRE(-RUB)(L)"      "RUB(L)"            "AL-DM6(R)"        
#> [148] "EPA(R)"            "AVLP(R)"           "AL-VP1d(R)"       
#> [151] "AB(L)"             "mALT(L)"           "PRW"              
#> [154] "b1(R)"             "VLNP(R)"           "IB"               
#> [157] "aL(R)"             "LX(R)"             "AL-VM5d(R)"       
#> [160] "SMP(R)"            "EBr6"              "a'1(R)"           
#> [163] "AL-D(L)"           "AME(R)"            "AL-DM5(R)"        
#> [166] "lACA(R)"           "NO1(L)"            "vACA(R)"          
#> [169] "PB"                "AL-VA1d(R)"        "AL-DM4(L)"        
#> [172] "PB(L9)"            "NO3(L)"            "AL-VA7m(R)"       
#> [175] "AL-VM7v(R)"        "AL-VP5(R)"         "AL-DM3(L)"        
#> [178] "PB(R1)"            "AL-DA2(L)"         "NO(R)"            
#> [181] "AL-VC2(R)"         "FB"                "ME(R)"            
#> [184] "b'2(R)"            "LAL(L)"            "MB(L)"            
#> [187] "AL-DA4m(L)"        "AL-DM5(L)"         "AL-DA4m(R)"       
#> [190] "AL-DP1m(R)"        "LAL(-GA)(R)"       "b2(R)"            
#> [193] "g5(R)"             "ICL(R)"            "AL-DC4(R)"        
#> [196] "EBr5"              "PB(R8)"            "AL-D(R)"          
#> [199] "SCL(R)"            "AL-DC1(R)"         "AL-DA4l(R)"       
#> [202] "AL-VM5v(R)"        "AL-DC1(L)"         "AL-DL4(R)"        
#> [205] "PB(R2)"            "AL-DL2v(R)"        "GA(R)"            
#> [208] "mALT(R)"           "FBl4"              "ATL(R)"           
#> [211] "dACA(R)"           "SNP(R)"            "PB(L6)"           
#> [214] "FBl1"              "a'L(L)"            "PB(R5)"           
#> [217] "EBr1"              "NO1(R)"            "AL-DM2(L)"        
#> [220] "aL(L)"             "AL-VA3(R)"         "AL-VC1(R)"        
#> [223] "SPS(L)"            "EB"                "PB(R7)"           
#> [226] "PLP(R)"            "bL(L)"             "AL-VP1l(R)"       
#> [229] "AOTU(R)"           "PB(R9)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "OL(R)"             "AL-DL4(R)"         "AL-VA7m(R)"       
#>   [4] "aL(R)"             "EB"                "VES(R)"           
#>   [7] "GNG"               "AL-VP1d(R)"        "SIP(L)"           
#>  [10] "AL-DM6(L)"         "b'L(L)"            "AL-VP4(R)"        
#>  [13] "AL-VP2(R)"         "ICL(L)"            "AB(R)"            
#>  [16] "AL-DA3(R)"         "AL-VA2(R)"         "EBr2r4"           
#>  [19] "FB-column3"        "WED(R)"            "EPA(R)"           
#>  [22] "LX(L)"             "AL-V(R)"           "PB(R2)"           
#>  [25] "CAN(R)"            "bL(R)"             "PB(R4)"           
#>  [28] "AL-VC5(R)"         "b2(R)"             "PB(L8)"           
#>  [31] "EBr3d"             "AL-VM7v(L)"        "NO3(L)"           
#>  [34] "PED(R)"            "MB(R)"             "AL-VC1(R)"        
#>  [37] "LOP(R)"            "SNP(R)"            "AL-DA4m(L)"       
#>  [40] "SCL(L)"            "mALT(L)"           "a1(R)"            
#>  [43] "PB(L3)"            "mALT(R)"           "NO1(L)"           
#>  [46] "AL(L)"             "AL-DL3(R)"         "ME(R)"            
#>  [49] "BU(R)"             "AL-DC3(R)"         "NO"               
#>  [52] "PB(L7)"            "AL-DC2(R)"         "SMP(L)"           
#>  [55] "FB"                "aL(L)"             "EBr6"             
#>  [58] "NO1(R)"            "gL(L)"             "PB(R5)"           
#>  [61] "EBr3am"            "LAL(-GA)(R)"       "a'L(R)"           
#>  [64] "AL-DA1(R)"         "VLNP(R)"           "AB(L)"            
#>  [67] "GOR(L)"            "EBr1"              "ROB(R)"           
#>  [70] "a'L(L)"            "AL-DM1(R)"         "LAL(L)"           
#>  [73] "FBl9"              "AL-VC2(R)"         "AL-DM5(L)"        
#>  [76] "EBr3pw"            "IPS(R)"            "AL-VM3(R)"        
#>  [79] "SCL(R)"            "CA(R)"             "SLP(R)"           
#>  [82] "AL-VM5v(R)"        "CX"                "g1(R)"            
#>  [85] "MB(L)"             "AL-DA4l(R)"        "AL-VL2p(R)"       
#>  [88] "IB"                "AL-DL5(L)"         "PB(R3)"           
#>  [91] "AL-VA6(R)"         "SPS(R)"            "SNP(L)"           
#>  [94] "ATL(R)"            "AL-DM3(L)"         "bL(L)"            
#>  [97] "AL-DP1m(L)"        "AL-VC3l(R)"        "CRE(-RUB)(L)"     
#> [100] "VES(L)"            "AL-VA1v(R)"        "AMMC"             
#> [103] "AL-DA2(L)"         "MB(+ACA)(R)"       "POC"              
#> [106] "PB(R7)"            "AL-DC2(L)"         "AL(R)"            
#> [109] "AL-DP1m(R)"        "b'1(R)"            "AL-VM7d(R)"       
#> [112] "AOT(R)"            "AL-DA3(L)"         "AL-DM1(L)"        
#> [115] "FBl8"              "a3(R)"             "NO(L)"            
#> [118] "GA(R)"             "PB"                "FBl3"             
#> [121] "b1(R)"             "AL-DM4(L)"         "a'3(R)"           
#> [124] "AL-D(R)"           "PB(R8)"            "PB(R1)"           
#> [127] "FBl4"              "LH(R)"             "AL-DM4(R)"        
#> [130] "PB(R9)"            "CRE(-ROB,-RUB)(R)" "vACA(R)"          
#> [133] "g2(R)"             "PB(L1)"            "PLP(R)"           
#> [136] "AL-VA7l(R)"        "AL-VP3(R)"         "SAD(-AMMC)"       
#> [139] "AL-DL5(R)"         "dACA(R)"           "LAL(R)"           
#> [142] "PB(L4)"            "AL-VC4(R)"         "AL-VM1(R)"        
#> [145] "FLA(R)"            "AL-VM7d(L)"        "g5(R)"            
#> [148] "INP"               "SAD"               "BU(L)"            
#> [151] "NO(R)"             "AL-DL1(R)"         "GC"               
#> [154] "SPS(L)"            "AME(R)"            "PB(L5)"           
#> [157] "a2(R)"             "PB(L6)"            "GOR(R)"           
#> [160] "AL-VA1d(R)"        "AL-VL2a(R)"        "hemibrain"        
#> [163] "AL-DC4(L)"         "AL-DL4(L)"         "AL-VA5(R)"        
#> [166] "AL-DL2v(R)"        "AL-DM2(R)"         "b'2(R)"           
#> [169] "NO2(R)"            "RUB(L)"            "SMP(R)"           
#> [172] "CA(L)"             "LO(R)"             "FBl5"             
#> [175] "AL-DC4(R)"         "AL-DM5(R)"         "AL-VA3(R)"        
#> [178] "FBl1"              "PENP"              "PRW"              
#> [181] "g4(R)"             "gL(R)"             "PVLP(R)"          
#> [184] "LX(R)"             "AL-DA4m(R)"        "AL-DP1l(R)"       
#> [187] "PB(L2)"            "EBr5"              "AL-DC1(R)"        
#> [190] "AL-VM5d(R)"        "GF(R)"             "b'L(R)"           
#> [193] "AL-DA2(R)"         "FBl2"              "g3(R)"            
#> [196] "PB(L9)"            "AL-DM2(L)"         "FBl7"             
#> [199] "AL-VA4(R)"         "ATL(L)"            "NO2(L)"           
#> [202] "RUB(R)"            "PB(R6)"            "AL-VA6(L)"        
#> [205] "FBl6"              "AVLP(R)"           "AL-VP1l(R)"       
#> [208] "CRE(R)"            "a'2(R)"            "AL-D(L)"          
#> [211] "AL-DM3(R)"         "AL-VM7v(R)"        "EPA(L)"           
#> [214] "AL-VM4(R)"         "AOTU(R)"           "lACA(R)"          
#> [217] "SIP(R)"            "CRE(L)"            "a'1(R)"           
#> [220] "VMNP"              "AL-DC1(L)"         "AL-VM2(R)"        
#> [223] "ICL(R)"            "AL-VL1(R)"         "AL-VP5(R)"        
#> [226] "AL-VP1m(R)"        "NO3(R)"            "AL-DL2d(R)"       
#> [229] "AL-DM6(R)"         "AL-VC3m(R)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                          
#> [1,] "POST" "/api/npexplorer/findneurons" 
#> [2,] "GET"  "/api/npexplorer/completeness"
#> [3,] "POST" "/api/v:ver/raw/cypher/cypher"
#> [4,] "POST" "/api/raw/cypher/transaction" 
#> [5,] "GET"  "/api/v:ver/dbmeta/database"  
#> [6,] "GET"  "/api/npexplorer/findneurons" 
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getFindNeurons-fm" 
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCompleteness-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getFindNeurons-fm" 

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }