Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "PB(L3)"            "b2(R)"             "a'L(L)"           
#>   [4] "AMMC"              "g4(R)"             "AL-DM5(L)"        
#>   [7] "AL-VP4(R)"         "CA(L)"             "SIP(L)"           
#>  [10] "a1(R)"             "ME(R)"             "AL-DA4l(R)"       
#>  [13] "AL-VC2(R)"         "AL-VM5v(R)"        "NO3(R)"           
#>  [16] "EBr5"              "AL-DL4(L)"         "PB(R2)"           
#>  [19] "INP"               "AL-DA3(R)"         "AL-VM4(R)"        
#>  [22] "mALT(R)"           "NO3(L)"            "VES(L)"           
#>  [25] "AL-DC2(R)"         "AL-VL1(R)"         "g1(R)"            
#>  [28] "LX(R)"             "CRE(L)"            "AL-VP1d(R)"       
#>  [31] "FBl9"              "POC"               "a'L(R)"           
#>  [34] "PB(R9)"            "SNP(L)"            "FBl1"             
#>  [37] "FBl4"              "FBl5"              "PB(R5)"           
#>  [40] "GOR(L)"            "EBr1"              "AL-DL4(R)"        
#>  [43] "GNG"               "AL-DC3(R)"         "LOP(R)"           
#>  [46] "vACA(R)"           "PLP(R)"            "PB(R1)"           
#>  [49] "FBl2"              "LAL(R)"            "AOTU(R)"          
#>  [52] "EBr3d"             "PB(L4)"            "FB"               
#>  [55] "PB"                "AL-DM1(L)"         "PVLP(R)"          
#>  [58] "CAN(R)"            "AOT(R)"            "a'2(R)"           
#>  [61] "AL-DP1m(L)"        "AL-VA2(R)"         "AL-VL2a(R)"       
#>  [64] "AL-VM7v(R)"        "EBr3pw"            "AL-DM3(R)"        
#>  [67] "AL-VA7m(R)"        "AL-DM2(R)"         "ATL(L)"           
#>  [70] "PB(L6)"            "PB(L9)"            "AL-VP2(R)"        
#>  [73] "NO1(R)"            "LO(R)"             "AL-VA4(R)"        
#>  [76] "PB(L1)"            "AL-DL3(R)"         "AL-DM1(R)"        
#>  [79] "SMP(R)"            "VES(R)"            "LX(L)"            
#>  [82] "CX"                "b'L(R)"            "PB(R4)"           
#>  [85] "AL-DL2v(R)"        "NO2(L)"            "AL-D(R)"          
#>  [88] "WED(R)"            "EBr6"              "AL-VM3(R)"        
#>  [91] "AL-DA2(L)"         "AL-DA1(R)"         "AL-VA3(R)"        
#>  [94] "SPS(L)"            "NO(R)"             "AL-VC3l(R)"       
#>  [97] "AB(L)"             "AL-VP1l(R)"        "LAL(L)"           
#> [100] "PB(L8)"            "AL-DC2(L)"         "MB(R)"            
#> [103] "AL-VC5(R)"         "AL-VP5(R)"         "AL-DA3(L)"        
#> [106] "AL-VC4(R)"         "VLNP(R)"           "b'2(R)"           
#> [109] "BU(L)"             "CA(R)"             "GC"               
#> [112] "AL-DA4m(L)"        "EPA(L)"            "g3(R)"            
#> [115] "SNP(R)"            "AL-VM2(R)"         "FBl6"             
#> [118] "PB(R3)"            "a3(R)"             "NO(L)"            
#> [121] "AL-DL2d(R)"        "AL-V(R)"           "GA(R)"            
#> [124] "AL-DM6(L)"         "AL-VM1(R)"         "aL(R)"            
#> [127] "AL-DM2(L)"         "AL-VC1(R)"         "SAD(-AMMC)"       
#> [130] "FB-column3"        "GOR(R)"            "SCL(L)"           
#> [133] "SMP(L)"            "g2(R)"             "MB(+ACA)(R)"      
#> [136] "aL(L)"             "AL-DM5(R)"         "SLP(R)"           
#> [139] "AL-VP3(R)"         "PENP"              "a2(R)"            
#> [142] "SCL(R)"            "ICL(L)"            "PB(L7)"           
#> [145] "OL(R)"             "hemibrain"         "ROB(R)"           
#> [148] "AL-DA2(R)"         "AL-VA5(R)"         "b'1(R)"           
#> [151] "mALT(L)"           "AL-DP1l(R)"        "gL(R)"            
#> [154] "FBl3"              "AVLP(R)"           "IPS(R)"           
#> [157] "bL(R)"             "AL-DM4(R)"         "GF(R)"            
#> [160] "PED(R)"            "AL(L)"             "CRE(-RUB)(L)"     
#> [163] "dACA(R)"           "bL(L)"             "PB(L2)"           
#> [166] "EB"                "AL-D(L)"           "AL-VA6(L)"        
#> [169] "CRE(R)"            "gL(L)"             "AL-VP1m(R)"       
#> [172] "CRE(-ROB,-RUB)(R)" "AL-VM7v(L)"        "NO1(L)"           
#> [175] "EPA(R)"            "AL-DC4(L)"         "AL-DM4(L)"        
#> [178] "g5(R)"             "AL-DA4m(R)"        "AL-DM6(R)"        
#> [181] "SIP(R)"            "IB112"             "AME(R)"           
#> [184] "AL-DC4(R)"         "AL-VA1v(R)"        "PB(R6)"           
#> [187] "AL-DC1(L)"         "PB(R8)"            "LH(R)"            
#> [190] "FBl8"              "EBr3am"            "a'3(R)"           
#> [193] "AL-VM7d(L)"        "ATL(R)"            "AL-VC3m(R)"       
#> [196] "AL-VL2p(R)"        "SPS(R)"            "EBr2r4"           
#> [199] "b1(R)"             "NO2(R)"            "AL-VA1d(R)"       
#> [202] "AL-VA6(R)"         "RUB(R)"            "FLA(R)"           
#> [205] "AL-DL5(L)"         "AB(R)"             "PRW"              
#> [208] "MB(L)"             "BU(R)"             "AL-DL1(R)"        
#> [211] "AL-VA7l(R)"        "IB"                "AL-DM3(L)"        
#> [214] "FBl7"              "b'L(L)"            "lACA(R)"          
#> [217] "AL-VM5d(R)"        "AL-VM7d(R)"        "AL-DC1(R)"        
#> [220] "SAD"               "LAL(-GA)(R)"       "RUB(L)"           
#> [223] "AL-DL5(R)"         "VMNP"              "NO"               
#> [226] "a'1(R)"            "PB(L5)"            "PB(R7)"           
#> [229] "ICL(R)"            "AL(R)"             "AL-DP1m(R)"       
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "CA(R)"             "GNG"               "LAL(-GA)(R)"      
#>   [4] "a'L(R)"            "FBl4"              "FBl6"             
#>   [7] "RUB(R)"            "SNP(R)"            "AME(R)"           
#>  [10] "aL(R)"             "PRW"               "GOR(L)"           
#>  [13] "PB(R9)"            "SAD(-AMMC)"        "PB(R2)"           
#>  [16] "POC"               "vACA(R)"           "b'2(R)"           
#>  [19] "ICL(R)"            "a'L(L)"            "VES(L)"           
#>  [22] "AB(L)"             "IB"                "SCL(L)"           
#>  [25] "AMMC"              "ATL(R)"            "MB(+ACA)(R)"      
#>  [28] "FBl9"              "EBr5"              "NO2(L)"           
#>  [31] "g1(R)"             "FLA(R)"            "bL(L)"            
#>  [34] "ICL(L)"            "PB(L7)"            "PB(R3)"           
#>  [37] "PB(R4)"            "VLNP(R)"           "FB"               
#>  [40] "EBr1"              "PB(L1)"            "LX(R)"            
#>  [43] "LOP(R)"            "g3(R)"             "ROB(R)"           
#>  [46] "WED(R)"            "FBl1"              "NO2(R)"           
#>  [49] "dACA(R)"           "EBr3am"            "PVLP(R)"          
#>  [52] "EPA(L)"            "INP"               "CX"               
#>  [55] "CA(L)"             "PLP(R)"            "NO"               
#>  [58] "PB(L6)"            "LO(R)"             "MB(R)"            
#>  [61] "PB(R5)"            "VMNP"              "b'L(R)"           
#>  [64] "BU(L)"             "PB(L9)"            "g5(R)"            
#>  [67] "AVLP(R)"           "RUB(L)"            "CRE(R)"           
#>  [70] "PB(L8)"            "a'1(R)"            "PB(L4)"           
#>  [73] "AOT(R)"            "SMP(R)"            "SAD"              
#>  [76] "a3(R)"             "EB"                "a2(R)"            
#>  [79] "AL(L)"             "SIP(R)"            "NO(R)"            
#>  [82] "b'L(L)"            "PB(L3)"            "SNP(L)"           
#>  [85] "FBl5"              "b'1(R)"            "EPA(R)"           
#>  [88] "g2(R)"             "GOR(R)"            "AL(R)"            
#>  [91] "a'3(R)"            "PB(L5)"            "mALT(R)"          
#>  [94] "EBr3d"             "EBr3pw"            "FBl3"             
#>  [97] "SMP(L)"            "PB(R8)"            "ME(R)"            
#> [100] "BU(R)"             "LAL(R)"            "lACA(R)"          
#> [103] "a'2(R)"            "aL(L)"             "NO(L)"            
#> [106] "mALT(L)"           "SIP(L)"            "EBr2r4"           
#> [109] "gL(R)"             "EBr6"              "PB"               
#> [112] "PB(R1)"            "FBl8"              "LX(L)"            
#> [115] "SLP(R)"            "AOTU(R)"           "GF(R)"            
#> [118] "VES(R)"            "OL(R)"             "PB(L2)"           
#> [121] "ATL(L)"            "LAL(L)"            "b2(R)"            
#> [124] "g4(R)"             "NO3(L)"            "AB(R)"            
#> [127] "PENP"              "NO3(R)"            "NO1(R)"           
#> [130] "FBl7"              "IPS(R)"            "GA(R)"            
#> [133] "MB(L)"             "FBl2"              "bL(R)"            
#> [136] "PB(R7)"            "SCL(R)"            "gL(L)"            
#> [139] "LH(R)"             "b1(R)"             "CRE(-ROB,-RUB)(R)"
#> [142] "PB(R6)"            "GC"                "NO1(L)"           
#> [145] "CAN(R)"            "hemibrain"         "PED(R)"           
#> [148] "SPS(L)"            "CRE(L)"            "a1(R)"            
#> [151] "SPS(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "EBr5"              "AL-DA4m(L)"        "AL-DM6(L)"        
#>   [4] "FB"                "LH(R)"             "AL-DP1m(R)"       
#>   [7] "PRW"               "AL-VA1d(R)"        "PB"               
#>  [10] "AL-DC1(L)"         "SAD(-AMMC)"        "CRE(-ROB,-RUB)(R)"
#>  [13] "EBr3d"             "FB-column3"        "AL-DA4m(R)"       
#>  [16] "AL-DC4(R)"         "AL-VM7v(L)"        "b1(R)"            
#>  [19] "AL-VC3m(R)"        "AL-VL2p(R)"        "dACA(R)"          
#>  [22] "PB(R3)"            "b'L(L)"            "AL-DM2(R)"        
#>  [25] "AL-VC5(R)"         "AL-VP3(R)"         "CX"               
#>  [28] "aL(R)"             "AL-VL2a(R)"        "VLNP(R)"          
#>  [31] "AL-DM1(L)"         "FBl9"              "AL-VA1v(R)"       
#>  [34] "IB"                "FBl2"              "NO(R)"            
#>  [37] "LO(R)"             "PB(R2)"            "PB(R4)"           
#>  [40] "AL-VM1(R)"         "AL-DC2(L)"         "GOR(L)"           
#>  [43] "FBl1"              "NO(L)"             "AL-DC1(R)"        
#>  [46] "AL-DL2d(R)"        "AL-DL4(R)"         "AL-DM1(R)"        
#>  [49] "AOTU(R)"           "AL-VM2(R)"         "AL-D(R)"          
#>  [52] "FBl7"              "AL-DA3(R)"         "AL-VA3(R)"        
#>  [55] "LOP(R)"            "vACA(R)"           "FBl3"             
#>  [58] "FLA(R)"            "AL-VP1m(R)"        "RUB(R)"           
#>  [61] "hemibrain"         "ICL(L)"            "PB(L8)"           
#>  [64] "AL-DC3(R)"         "AL-VA7m(R)"        "SCL(R)"           
#>  [67] "PED(R)"            "CAN(R)"            "b'2(R)"           
#>  [70] "PB(L4)"            "a'L(L)"            "AL-VC3l(R)"       
#>  [73] "POC"               "GC"                "EBr3am"           
#>  [76] "GOR(R)"            "mALT(L)"           "mALT(R)"          
#>  [79] "MB(L)"             "ROB(R)"            "AL-DM6(R)"        
#>  [82] "NO2(L)"            "PB(R7)"            "AL-VC4(R)"        
#>  [85] "PVLP(R)"           "GF(R)"             "a'2(R)"           
#>  [88] "g5(R)"             "SNP(L)"            "AB(L)"            
#>  [91] "PB(L3)"            "PB(L1)"            "ME(R)"            
#>  [94] "g4(R)"             "AL-VL1(R)"         "SAD"              
#>  [97] "NO1(R)"            "BU(R)"             "AL-VM5d(R)"       
#> [100] "SPS(R)"            "LX(R)"             "AL-DP1m(L)"       
#> [103] "CA(R)"             "AL-DL5(R)"         "AL-DM5(R)"        
#> [106] "AL-VP1l(R)"        "EBr6"              "CRE(-RUB)(L)"     
#> [109] "ATL(R)"            "b2(R)"             "PB(L6)"           
#> [112] "AL-DA2(R)"         "AL-DA1(R)"         "PB(L9)"           
#> [115] "EBr3pw"            "VES(R)"            "LX(L)"            
#> [118] "AL-DM2(L)"         "g2(R)"             "AL-VM7d(L)"       
#> [121] "AL-VA5(R)"         "AL-VP2(R)"         "b'1(R)"           
#> [124] "LAL(L)"            "VES(L)"            "AL-V(R)"          
#> [127] "SCL(L)"            "PB(R8)"            "PB(R6)"           
#> [130] "BU(L)"             "AL-DL4(L)"         "FBl8"             
#> [133] "NO1(L)"            "lACA(R)"           "WED(R)"           
#> [136] "AL-DL3(R)"         "AL-VP5(R)"         "a'3(R)"           
#> [139] "SIP(L)"            "EB"                "PB(R1)"           
#> [142] "gL(R)"             "ICL(R)"            "AB(R)"            
#> [145] "AL-VC1(R)"         "SMP(R)"            "g1(R)"            
#> [148] "AL-DL2v(R)"        "CRE(L)"            "g3(R)"            
#> [151] "AL-DM4(L)"         "AL-DM3(L)"         "a2(R)"            
#> [154] "PB(R9)"            "IPS(R)"            "NO2(R)"           
#> [157] "PB(L7)"            "SPS(L)"            "SMP(L)"           
#> [160] "AL-DA2(L)"         "MB(+ACA)(R)"       "AL-VM3(R)"        
#> [163] "AMMC"              "a'L(R)"            "PB(L5)"           
#> [166] "PB(R5)"            "FBl4"              "AL-DP1l(R)"       
#> [169] "AL-VA4(R)"         "AL-VM4(R)"         "AL-DM5(L)"        
#> [172] "EBr2r4"            "aL(L)"             "AL-DL1(R)"        
#> [175] "AL-DM3(R)"         "AL-VA6(R)"         "AL-VP1d(R)"       
#> [178] "VMNP"              "AL(R)"             "bL(L)"            
#> [181] "AL-VP4(R)"         "AL-DC2(R)"         "bL(R)"            
#> [184] "OL(R)"             "EBr1"              "AL-VA7l(R)"       
#> [187] "gL(L)"             "EPA(L)"            "SLP(R)"           
#> [190] "AL-VA2(R)"         "a1(R)"             "SNP(R)"           
#> [193] "AL-DA3(L)"         "NO"                "PB(L2)"           
#> [196] "CA(L)"             "CRE(R)"            "PLP(R)"           
#> [199] "EPA(R)"            "a'1(R)"            "INP"              
#> [202] "PENP"              "a3(R)"             "NO3(R)"           
#> [205] "AL-VA6(L)"         "AME(R)"            "AL-DA4l(R)"       
#> [208] "AL-VC2(R)"         "AL-D(L)"           "b'L(R)"           
#> [211] "SIP(R)"            "FBl6"              "AVLP(R)"          
#> [214] "RUB(L)"            "AL-VM5v(R)"        "LAL(-GA)(R)"      
#> [217] "AL-DC4(L)"         "AL-DL5(L)"         "NO3(L)"           
#> [220] "FBl5"              "MB(R)"             "GA(R)"            
#> [223] "GNG"               "AL-DM4(R)"         "AL-VM7d(R)"       
#> [226] "AOT(R)"            "ATL(L)"            "AL(L)"            
#> [229] "AL-VM7v(R)"        "LAL(R)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "LAL(R)"            "AL-VA3(R)"         "AL-VP3(R)"        
#>   [4] "AOTU(R)"           "FLA(R)"            "PB(R2)"           
#>   [7] "AL-VM7v(R)"        "SPS(R)"            "AL-D(L)"          
#>  [10] "AL-DL5(R)"         "LH(R)"             "AL-DA4m(R)"       
#>  [13] "AL-DM5(R)"         "AL-DP1m(R)"        "LAL(-GA)(R)"      
#>  [16] "AL-DC2(L)"         "WED(R)"            "PB(L1)"           
#>  [19] "FBl1"              "NO(L)"             "b2(R)"            
#>  [22] "b'1(R)"            "FBl9"              "PB(R3)"           
#>  [25] "AL-VM3(R)"         "GA(R)"             "MB(L)"            
#>  [28] "BU(R)"             "AL-DM6(L)"         "SCL(R)"           
#>  [31] "EBr3am"            "AL-VM5d(R)"        "PB(L2)"           
#>  [34] "EPA(R)"            "AL-DM2(L)"         "bL(R)"            
#>  [37] "AL-VC3m(R)"        "FBl6"              "AL-VP1l(R)"       
#>  [40] "SIP(L)"            "g4(R)"             "AL-DC4(R)"        
#>  [43] "PB"                "EBr3d"             "ICL(L)"           
#>  [46] "aL(L)"             "AL-DA3(R)"         "LOP(R)"           
#>  [49] "SNP(R)"            "ROB(R)"            "CRE(L)"           
#>  [52] "AME(R)"            "EBr2r4"            "AL-VM7d(L)"       
#>  [55] "CRE(-RUB)(L)"      "CA(L)"             "VLNP(R)"          
#>  [58] "b'2(R)"            "SPS(L)"            "NO3(L)"           
#>  [61] "AL-VP5(R)"         "b'L(L)"            "AL-DA1(R)"        
#>  [64] "AL-DP1l(R)"        "g2(R)"             "MB(+ACA)(R)"      
#>  [67] "AL-VP1d(R)"        "NO2(L)"            "SIP(R)"           
#>  [70] "PB(R4)"            "GOR(R)"            "AL-DL3(R)"        
#>  [73] "AL-VP2(R)"         "GC"                "VMNP"             
#>  [76] "FBl2"              "AL-VC1(R)"         "FBl7"             
#>  [79] "AL(R)"             "AL-DA2(R)"         "AL-DA4l(R)"       
#>  [82] "AL-VC2(R)"         "PENP"              "PB(R6)"           
#>  [85] "AL-DM1(L)"         "PB(L9)"            "CAN(R)"           
#>  [88] "PB(L4)"            "AL-DC2(R)"         "AL-VM5v(R)"       
#>  [91] "a1(R)"             "SAD(-AMMC)"        "AVLP(R)"          
#>  [94] "PB(R8)"            "bL(L)"             "EBr3pw"           
#>  [97] "SNP(L)"            "AL-VP1m(R)"        "CRE(R)"           
#> [100] "AL-DC1(L)"         "ATL(R)"            "AL-VM2(R)"        
#> [103] "SAD"               "b'L(R)"            "SLP(R)"           
#> [106] "AL-VC4(R)"         "VES(R)"            "AL-DL2d(R)"       
#> [109] "AL-DM1(R)"         "AL-DM2(R)"         "a'L(L)"           
#> [112] "AL-D(R)"           "mALT(L)"           "PLP(R)"           
#> [115] "ME(R)"             "AL-DA2(L)"         "FBl8"             
#> [118] "AL-VM1(R)"         "AL-VA5(R)"         "AB(L)"            
#> [121] "NO1(L)"            "ICL(R)"            "VES(L)"           
#> [124] "AL-DM4(R)"         "mALT(R)"           "AL-VL2p(R)"       
#> [127] "AL-VA7l(R)"        "vACA(R)"           "GF(R)"            
#> [130] "PB(L8)"            "AL-DM3(L)"         "EPA(L)"           
#> [133] "AL-DA3(L)"         "CA(R)"             "GOR(L)"           
#> [136] "AL-VA4(R)"         "AL-VM4(R)"         "PED(R)"           
#> [139] "FB"                "NO"                "PB(L3)"           
#> [142] "AL-DL1(R)"         "CRE(-ROB,-RUB)(R)" "ATL(L)"           
#> [145] "g1(R)"             "a'1(R)"            "LX(L)"            
#> [148] "GNG"               "AL-DC4(L)"         "MB(R)"            
#> [151] "FB-column3"        "AL-VA1v(R)"        "a'3(R)"           
#> [154] "AL-DL4(R)"         "AOT(R)"            "AL-DL4(L)"        
#> [157] "FBl5"              "INP"               "AL-DC1(R)"        
#> [160] "RUB(R)"            "PRW"               "EBr6"             
#> [163] "a'L(R)"            "AL-DM6(R)"         "g3(R)"            
#> [166] "EBr5"              "AL-VC5(R)"         "gL(R)"            
#> [169] "AL-VL2a(R)"        "EBr1"              "AL-DL5(L)"        
#> [172] "PB(R9)"            "PB(R1)"            "AL-VM7v(L)"       
#> [175] "PB(R5)"            "PB(L5)"            "AMMC"             
#> [178] "LAL(L)"            "AL-DM5(L)"         "IPS(R)"           
#> [181] "RUB(L)"            "AB(R)"             "PB(R7)"           
#> [184] "a'2(R)"            "AL-DP1m(L)"        "g5(R)"            
#> [187] "NO(R)"             "IB"                "AL-VA7m(R)"       
#> [190] "AL-V(R)"           "dACA(R)"           "NO1(R)"           
#> [193] "LX(R)"             "AL-DM4(L)"         "OL(R)"            
#> [196] "a2(R)"             "LO(R)"             "b1(R)"            
#> [199] "AL-VM7d(R)"        "AL-VL1(R)"         "EB"               
#> [202] "AL-DA4m(L)"        "PB(L7)"            "SMP(L)"           
#> [205] "AL-VA6(L)"         "PVLP(R)"           "CX"               
#> [208] "aL(R)"             "gL(L)"             "AL-VA2(R)"        
#> [211] "POC"               "hemibrain"         "NO3(R)"           
#> [214] "SMP(R)"            "AL-DL2v(R)"        "AL(L)"            
#> [217] "FBl3"              "FBl4"              "AL-VP4(R)"        
#> [220] "AL-DM3(R)"         "AL-VC3l(R)"        "PB(L6)"           
#> [223] "a3(R)"             "AL-DC3(R)"         "AL-VA6(R)"        
#> [226] "SCL(L)"            "lACA(R)"           "BU(L)"            
#> [229] "NO2(R)"            "AL-VA1d(R)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method     path                                   
#> [1,] "HEAD"     "/api"                                 
#> [2,] "PROPFIND" "/api/*"                               
#> [3,] "GET"      "/api/v:ver/npexplorer/roiconnectivity"
#> [4,] "GET"      "/api/v:ver/npexplorer/rankedtable"    
#> [5,] "GET"      "/api/v:ver/dbmeta/instances"          
#> [6,] "POST"     "/api/npexplorer/roiconnectivity"      
#>      name                                                                                           
#> [1,] "github.com/labstack/echo.glob..func1"                                                         
#> [2,] "github.com/labstack/echo.glob..func1"                                                         
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIConnectivity-fm"
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }