Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "RUB(L)"            "AL-VC4(R)"         "SNP(R)"           
#>   [4] "AL-DL5(L)"         "AL-DL5(R)"         "AL-V(R)"          
#>   [7] "IB"                "EBr6"              "PLP(R)"           
#>  [10] "NO(R)"             "LO(R)"             "AL-VA1d(R)"       
#>  [13] "b'1(R)"            "AL-DA3(L)"         "AL-DM4(L)"        
#>  [16] "b'L(R)"            "AVLP(R)"           "VMNP"             
#>  [19] "SCL(L)"            "SMP(R)"            "CRE(-ROB,-RUB)(R)"
#>  [22] "g4(R)"             "AL-DA3(R)"         "AL-DC4(R)"        
#>  [25] "AL-DM3(R)"         "AL-VA7l(R)"        "AL-VL1(R)"        
#>  [28] "AL-VL2a(R)"        "EBr3d"             "AL-DM5(L)"        
#>  [31] "a'1(R)"            "PRW"               "gL(L)"            
#>  [34] "BU(R)"             "AL-DM3(L)"         "SIP(R)"           
#>  [37] "FBl1"              "AL-DP1m(R)"        "AL-VP1l(R)"       
#>  [40] "CRE(R)"            "a'2(R)"            "g5(R)"            
#>  [43] "AL-VA2(R)"         "PB(R7)"            "PB(L6)"           
#>  [46] "MB(R)"             "lACA(R)"           "NO1(R)"           
#>  [49] "AL-VA6(R)"         "a'L(R)"            "AL-DM4(R)"        
#>  [52] "PB(R2)"            "AL-VM5v(R)"        "PB(R5)"           
#>  [55] "g3(R)"             "NO1(L)"            "LH(R)"            
#>  [58] "AB(R)"             "bL(R)"             "AL-DM5(R)"        
#>  [61] "AL-VL2p(R)"        "EBr1"              "FBl4"             
#>  [64] "PB(L9)"            "AL-DC1(R)"         "vACA(R)"          
#>  [67] "GF(R)"             "SIP(L)"            "WED(R)"           
#>  [70] "FBl2"              "AL-VP1d(R)"        "AL-DM1(L)"        
#>  [73] "INP"               "AL-DA4m(L)"        "AL-DM2(R)"        
#>  [76] "FLA(R)"            "ME(R)"             "CRE(-RUB)(L)"     
#>  [79] "b1(R)"             "PB(R4)"            "BU(L)"            
#>  [82] "PB"                "AL-VA1v(R)"        "CX"               
#>  [85] "a2(R)"             "AME(R)"            "aL(L)"            
#>  [88] "AL-DL4(R)"         "AL-VC3l(R)"        "a1(R)"            
#>  [91] "GOR(L)"            "PB(L7)"            "a3(R)"            
#>  [94] "AL-VC5(R)"         "CA(L)"             "LX(R)"            
#>  [97] "AL-DC2(L)"         "AL-D(R)"           "EPA(R)"           
#> [100] "AL-VM7v(L)"        "AL-VA6(L)"         "AL-VM3(R)"        
#> [103] "FBl6"              "MB(L)"             "PVLP(R)"          
#> [106] "mALT(L)"           "PB(R6)"            "CA(R)"            
#> [109] "PB(L2)"            "LAL(-GA)(R)"       "LX(L)"            
#> [112] "FBl9"              "AL-DM2(L)"         "PB(R1)"           
#> [115] "AL-DA2(R)"         "AL-DM1(R)"         "AL-VM7v(R)"       
#> [118] "PB(L3)"            "PB(L1)"            "AL-DA4m(R)"       
#> [121] "b2(R)"             "b'L(L)"            "FB-column3"       
#> [124] "SPS(R)"            "IB112"             "AL-DM6(L)"        
#> [127] "NO3(L)"            "AL-VA4(R)"         "AL-VM5d(R)"       
#> [130] "PB(L8)"            "AL-DL3(R)"         "AL-DA2(L)"        
#> [133] "AL-DC4(L)"         "CAN(R)"            "AL-VP5(R)"        
#> [136] "GA(R)"             "hemibrain"         "NO(L)"            
#> [139] "AL-VA7m(R)"        "gL(R)"             "CRE(L)"           
#> [142] "FBl7"              "AL-DL2v(R)"        "AL-VM1(R)"        
#> [145] "dACA(R)"           "PB(R8)"            "EBr5"             
#> [148] "ICL(L)"            "EPA(L)"            "MB(+ACA)(R)"      
#> [151] "FBl8"              "EB"                "AL-VM7d(L)"       
#> [154] "a'L(L)"            "ATL(R)"            "AL-DA1(R)"        
#> [157] "AMMC"              "SMP(L)"            "SPS(L)"           
#> [160] "AL(R)"             "AL-VC2(R)"         "FB"               
#> [163] "OL(R)"             "AL-VA3(R)"         "FBl5"             
#> [166] "LAL(R)"            "VES(L)"            "AL-VA5(R)"        
#> [169] "EBr2r4"            "POC"               "AOT(R)"           
#> [172] "FBl3"              "NO2(R)"            "AL-VP3(R)"        
#> [175] "a'3(R)"            "AL-DL4(L)"         "AL-D(L)"          
#> [178] "ATL(L)"            "AL(L)"             "AL-DL2d(R)"       
#> [181] "NO2(L)"            "AL-VC1(R)"         "AL-VM2(R)"        
#> [184] "PED(R)"            "AL-VP2(R)"         "AL-DM6(R)"        
#> [187] "AL-VP1m(R)"        "AB(L)"             "RUB(R)"           
#> [190] "AL-DP1m(L)"        "PB(R9)"            "SCL(R)"           
#> [193] "SLP(R)"            "AL-DP1l(R)"        "AL-VC3m(R)"       
#> [196] "AL-VM7d(R)"        "mALT(R)"           "GC"               
#> [199] "SAD(-AMMC)"        "AL-DC3(R)"         "AL-DL1(R)"        
#> [202] "PB(L5)"            "bL(L)"             "SNP(L)"           
#> [205] "LOP(R)"            "PB(L4)"            "EBr3am"           
#> [208] "GNG"               "AL-VM4(R)"         "PENP"             
#> [211] "PB(R3)"            "VLNP(R)"           "AOTU(R)"          
#> [214] "NO3(R)"            "NO"                "VES(R)"           
#> [217] "GOR(R)"            "SAD"               "g1(R)"            
#> [220] "AL-VP4(R)"         "ROB(R)"            "AL-DA4l(R)"       
#> [223] "g2(R)"             "ICL(R)"            "IPS(R)"           
#> [226] "LAL(L)"            "aL(R)"             "EBr3pw"           
#> [229] "AL-DC2(R)"         "b'2(R)"            "AL-DC1(L)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "BU(L)"             "ICL(L)"            "b'L(R)"           
#>   [4] "AL(L)"             "PB(R9)"            "EB"               
#>   [7] "WED(R)"            "FLA(R)"            "LX(R)"            
#>  [10] "g4(R)"             "PB(L9)"            "PVLP(R)"          
#>  [13] "GNG"               "PB(R7)"            "SIP(R)"           
#>  [16] "PB(L6)"            "MB(+ACA)(R)"       "PB(L4)"           
#>  [19] "CRE(-ROB,-RUB)(R)" "dACA(R)"           "EBr5"             
#>  [22] "EBr3pw"            "LOP(R)"            "hemibrain"        
#>  [25] "OL(R)"             "AB(L)"             "EPA(L)"           
#>  [28] "FBl7"              "ICL(R)"            "PB(L2)"           
#>  [31] "SPS(R)"            "IB"                "IPS(R)"           
#>  [34] "ROB(R)"            "FBl1"              "a3(R)"            
#>  [37] "RUB(R)"            "mALT(L)"           "a1(R)"            
#>  [40] "EBr3d"             "SMP(L)"            "g5(R)"            
#>  [43] "LH(R)"             "bL(R)"             "EBr3am"           
#>  [46] "GF(R)"             "NO1(L)"            "SPS(L)"           
#>  [49] "PLP(R)"            "NO(R)"             "AOT(R)"           
#>  [52] "SAD"               "gL(R)"             "FBl2"             
#>  [55] "MB(R)"             "lACA(R)"           "NO3(R)"           
#>  [58] "MB(L)"             "gL(L)"             "PB(L7)"           
#>  [61] "CAN(R)"            "AMMC"              "LAL(L)"           
#>  [64] "EBr6"              "a'2(R)"            "LX(L)"            
#>  [67] "GA(R)"             "NO3(L)"            "aL(L)"            
#>  [70] "RUB(L)"            "SMP(R)"            "GOR(L)"           
#>  [73] "BU(R)"             "PB(R1)"            "GOR(R)"           
#>  [76] "NO(L)"             "b'1(R)"            "a2(R)"            
#>  [79] "PB(L1)"            "PB(R6)"            "PB(L8)"           
#>  [82] "ME(R)"             "NO"                "VES(R)"           
#>  [85] "FBl3"              "ATL(R)"            "VLNP(R)"          
#>  [88] "aL(R)"             "mALT(R)"           "EPA(R)"           
#>  [91] "a'1(R)"            "bL(L)"             "SNP(L)"           
#>  [94] "PED(R)"            "SCL(L)"            "PRW"              
#>  [97] "GC"                "CA(R)"             "b1(R)"            
#> [100] "PB(R4)"            "vACA(R)"           "AOTU(R)"          
#> [103] "SCL(R)"            "AB(R)"             "g1(R)"            
#> [106] "SLP(R)"            "CX"                "a'L(L)"           
#> [109] "NO1(R)"            "VMNP"              "LAL(-GA)(R)"      
#> [112] "CRE(R)"            "FB"                "g2(R)"            
#> [115] "LO(R)"             "FBl4"              "PB(R5)"           
#> [118] "CA(L)"             "PB(L3)"            "INP"              
#> [121] "AVLP(R)"           "LAL(R)"            "VES(L)"           
#> [124] "POC"               "a'3(R)"            "FBl5"             
#> [127] "EBr1"              "PB(R8)"            "a'L(R)"           
#> [130] "FBl9"              "PB(R2)"            "AL(R)"            
#> [133] "PB(L5)"            "g3(R)"             "ATL(L)"           
#> [136] "EBr2r4"            "CRE(L)"            "FBl6"             
#> [139] "PB(R3)"            "AME(R)"            "NO2(R)"           
#> [142] "b'L(L)"            "PENP"              "NO2(L)"           
#> [145] "SAD(-AMMC)"        "FBl8"              "b2(R)"            
#> [148] "SIP(L)"            "PB"                "SNP(R)"           
#> [151] "b'2(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "PB(R2)"            "AL-VA1d(R)"        "AL-VM1(R)"        
#>   [4] "AL-VM3(R)"         "LAL(-GA)(R)"       "g1(R)"            
#>   [7] "FLA(R)"            "AL-DC2(L)"         "gL(R)"            
#>  [10] "AMMC"              "PENP"              "EBr1"             
#>  [13] "AL-DC1(L)"         "AL-DP1l(R)"        "EPA(R)"           
#>  [16] "AL-DC4(L)"         "MB(+ACA)(R)"       "AB(R)"            
#>  [19] "LAL(R)"            "AL-VP4(R)"         "LOP(R)"           
#>  [22] "AL-DM5(R)"         "CX"                "PB(L8)"           
#>  [25] "AL-DL5(R)"         "AL-VP1m(R)"        "ATL(L)"           
#>  [28] "PB(R4)"            "AL-V(R)"           "b'1(R)"           
#>  [31] "aL(R)"             "FBl9"              "AL-VC2(R)"        
#>  [34] "PB(L5)"            "BU(L)"             "AL-DM6(L)"        
#>  [37] "AL-VM7d(L)"        "AL-VC3m(R)"        "SMP(R)"           
#>  [40] "NO1(L)"            "AL-DA2(L)"         "AL-VA3(R)"        
#>  [43] "EBr6"              "AL-VA2(R)"         "AL-VA7l(R)"       
#>  [46] "g3(R)"             "GA(R)"             "ICL(L)"           
#>  [49] "CRE(R)"            "NO"                "CRE(L)"           
#>  [52] "AL-DM6(R)"         "AL-VP1d(R)"        "a'L(R)"           
#>  [55] "SIP(R)"            "FBl3"              "CA(R)"            
#>  [58] "a2(R)"             "b'L(R)"            "SNP(R)"           
#>  [61] "PB(L2)"            "GOR(L)"            "AL-DM5(L)"        
#>  [64] "AL-DC4(R)"         "AL-VA1v(R)"        "AL-VM5d(R)"       
#>  [67] "NO3(R)"            "AL-DA4m(L)"        "ICL(R)"           
#>  [70] "AOT(R)"            "PB(R5)"            "mALT(L)"          
#>  [73] "LAL(L)"            "PB(R3)"            "AL-VL2a(R)"       
#>  [76] "NO2(R)"            "RUB(L)"            "AB(L)"            
#>  [79] "SMP(L)"            "PB"                "AL-DM1(L)"        
#>  [82] "LX(L)"             "PED(R)"            "EBr5"             
#>  [85] "b1(R)"             "SPS(L)"            "a3(R)"            
#>  [88] "AL-VA7m(R)"        "AL-DL2d(R)"        "VLNP(R)"          
#>  [91] "AL-VC1(R)"         "EBr2r4"            "b'L(L)"           
#>  [94] "AL-DP1m(R)"        "ATL(R)"            "AL-DC1(R)"        
#>  [97] "AL-VP3(R)"         "ME(R)"             "NO3(L)"           
#> [100] "IPS(R)"            "AL-VM4(R)"         "SCL(L)"           
#> [103] "CA(L)"             "AL-DP1m(L)"        "LO(R)"            
#> [106] "GOR(R)"            "SAD"               "FBl8"             
#> [109] "AL-DM2(R)"         "AL-DA1(R)"         "GF(R)"            
#> [112] "EB"                "FBl1"              "SCL(R)"           
#> [115] "AL-VP1l(R)"        "AL-VP2(R)"         "PB(L6)"           
#> [118] "PB(L4)"            "AL-VM5v(R)"        "AL-DA3(L)"        
#> [121] "AL-DA2(R)"         "EPA(L)"            "AL-DL1(R)"        
#> [124] "AL-DM2(L)"         "AL-DL2v(R)"        "RUB(R)"           
#> [127] "PB(R7)"            "VMNP"              "AL-DL5(L)"        
#> [130] "PB(L7)"            "AL-DC3(R)"         "AL-DM1(R)"        
#> [133] "a'1(R)"            "g5(R)"             "FBl5"             
#> [136] "AL-VM7d(R)"        "AL-VP5(R)"         "AL(L)"            
#> [139] "EBr3pw"            "SAD(-AMMC)"        "AL-DL4(L)"        
#> [142] "AME(R)"            "AL-DL4(R)"         "EBr3d"            
#> [145] "AL(R)"             "AL-VC3l(R)"        "PB(L1)"           
#> [148] "NO1(R)"            "AL-DA4m(R)"        "gL(L)"            
#> [151] "OL(R)"             "AL-VA6(L)"         "FBl6"             
#> [154] "AL-DA3(R)"         "AL-VC5(R)"         "hemibrain"        
#> [157] "lACA(R)"           "MB(L)"             "BU(R)"            
#> [160] "MB(R)"             "a'3(R)"            "SPS(R)"           
#> [163] "AL-DL3(R)"         "dACA(R)"           "AL-D(L)"          
#> [166] "FBl7"              "AL-VM7v(L)"        "PB(R9)"           
#> [169] "AL-DM3(R)"         "AL-VC4(R)"         "AL-VL2p(R)"       
#> [172] "WED(R)"            "a'2(R)"            "AL-DM4(L)"        
#> [175] "AL-VM7v(R)"        "vACA(R)"           "a1(R)"            
#> [178] "bL(R)"             "a'L(L)"            "GC"               
#> [181] "g2(R)"             "bL(L)"             "AL-D(R)"          
#> [184] "AL-VA6(R)"         "IB"                "EBr3am"           
#> [187] "g4(R)"             "ROB(R)"            "FBl4"             
#> [190] "NO(L)"             "AL-VA5(R)"         "PB(R6)"           
#> [193] "b'2(R)"            "INP"               "PB(R1)"           
#> [196] "FB-column3"        "NO2(L)"            "FBl2"             
#> [199] "VES(L)"            "SLP(R)"            "AL-DM4(R)"        
#> [202] "AL-VA4(R)"         "b2(R)"             "AL-DM3(L)"        
#> [205] "PB(L9)"            "AL-VL1(R)"         "AL-VM2(R)"        
#> [208] "FB"                "AL-DA4l(R)"        "PB(L3)"           
#> [211] "PLP(R)"            "aL(L)"             "POC"              
#> [214] "AOTU(R)"           "PB(R8)"            "CAN(R)"           
#> [217] "PVLP(R)"           "LX(R)"             "NO(R)"            
#> [220] "CRE(-RUB)(L)"      "mALT(R)"           "GNG"              
#> [223] "CRE(-ROB,-RUB)(R)" "AL-DC2(R)"         "SIP(L)"           
#> [226] "SNP(L)"            "LH(R)"             "PRW"              
#> [229] "VES(R)"            "AVLP(R)"          
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "PB(R6)"            "AL-VM7v(L)"        "LAL(-GA)(R)"      
#>   [4] "b'L(R)"            "AL-DM4(R)"         "b'1(R)"           
#>   [7] "AL-DL5(R)"         "AL-DP1m(R)"        "AL-VA2(R)"        
#>  [10] "LAL(L)"            "GA(R)"             "AL-DA3(L)"        
#>  [13] "AL-DA1(R)"         "PB"                "AL-DL4(R)"        
#>  [16] "AL-VA7m(R)"        "ATL(L)"            "PB(R8)"           
#>  [19] "AL-DL5(L)"         "PB(L2)"            "WED(R)"           
#>  [22] "NO3(R)"            "PB(L7)"            "AL-VA4(R)"        
#>  [25] "AL-DP1m(L)"        "AL-VC3l(R)"        "SCL(L)"           
#>  [28] "b'L(L)"            "BU(L)"             "LX(R)"            
#>  [31] "CRE(L)"            "NO2(R)"            "AL-DM6(L)"        
#>  [34] "AL-DL2v(R)"        "AL-VM7v(R)"        "LOP(R)"           
#>  [37] "a'3(R)"            "NO1(R)"            "FBl5"             
#>  [40] "FBl8"              "AL-VL1(R)"         "CA(L)"            
#>  [43] "EBr5"              "PLP(R)"            "PB(R9)"           
#>  [46] "CRE(-RUB)(L)"      "GF(R)"             "dACA(R)"          
#>  [49] "g5(R)"             "AL-VL2p(R)"        "AB(L)"            
#>  [52] "NO1(L)"            "NO(R)"             "SPS(L)"           
#>  [55] "LO(R)"             "AL-DL2d(R)"        "SMP(R)"           
#>  [58] "CRE(-ROB,-RUB)(R)" "AL-DC1(L)"         "LH(R)"            
#>  [61] "AL-D(R)"           "AL-VP3(R)"         "AL-VM3(R)"        
#>  [64] "SMP(L)"            "EBr1"              "VES(L)"           
#>  [67] "AL-DM3(R)"         "AVLP(R)"           "CAN(R)"           
#>  [70] "AL-VL2a(R)"        "PB(R5)"            "MB(L)"            
#>  [73] "a'L(R)"            "PB(L9)"            "ICL(L)"           
#>  [76] "vACA(R)"           "PVLP(R)"           "AL-DC3(R)"        
#>  [79] "AL-VM2(R)"         "AL-VP2(R)"         "SAD"              
#>  [82] "SIP(L)"            "AL-D(L)"           "AL-VA6(L)"        
#>  [85] "PB(L3)"            "AOTU(R)"           "FBl3"             
#>  [88] "g1(R)"             "AL-DM3(L)"         "FLA(R)"           
#>  [91] "AL-VM5d(R)"        "a3(R)"             "AL-VC1(R)"        
#>  [94] "PRW"               "AL(L)"             "LX(L)"            
#>  [97] "INP"               "a2(R)"             "MB(R)"            
#> [100] "AL-VA7l(R)"        "FB-column3"        "PB(R2)"           
#> [103] "AL-V(R)"           "AMMC"              "CRE(R)"           
#> [106] "ME(R)"             "AME(R)"            "AL-VP5(R)"        
#> [109] "POC"               "aL(R)"             "SIP(R)"           
#> [112] "a'1(R)"            "a'L(L)"            "PB(R4)"           
#> [115] "GNG"               "AL-DC1(R)"         "PB(L8)"           
#> [118] "IPS(R)"            "SLP(R)"            "AL-VM4(R)"        
#> [121] "b'2(R)"            "AL-DM4(L)"         "ROB(R)"           
#> [124] "PB(L6)"            "IB"                "NO2(L)"           
#> [127] "AL-DA2(L)"         "FBl7"              "CA(R)"            
#> [130] "AL(R)"             "mALT(R)"           "ICL(R)"           
#> [133] "AL-VP1m(R)"        "a1(R)"             "AL-VP1d(R)"       
#> [136] "PB(R7)"            "aL(L)"             "AL-DM1(R)"        
#> [139] "AL-VA1d(R)"        "gL(L)"             "g4(R)"            
#> [142] "AL-VC2(R)"         "VLNP(R)"           "AL-VM7d(L)"       
#> [145] "LAL(R)"            "AL-DL3(R)"         "AL-DM2(R)"        
#> [148] "gL(R)"             "bL(L)"             "AL-VA1v(R)"       
#> [151] "AOT(R)"            "AL-DC2(L)"         "FBl4"             
#> [154] "EPA(L)"            "RUB(L)"            "EB"               
#> [157] "AL-DL4(L)"         "NO"                "AL-DA3(R)"        
#> [160] "EBr3am"            "SPS(R)"            "AL-DM2(L)"        
#> [163] "NO(L)"             "AL-DM6(R)"         "GC"               
#> [166] "FB"                "SNP(R)"            "FBl1"             
#> [169] "CX"                "SNP(L)"            "SAD(-AMMC)"       
#> [172] "AL-DC4(R)"         "AL-VC5(R)"         "EBr2r4"           
#> [175] "AL-DA4m(L)"        "PB(L5)"            "AL-VM5v(R)"       
#> [178] "RUB(R)"            "AL-VM1(R)"         "AL-VP1l(R)"       
#> [181] "g2(R)"             "PB(R1)"            "FBl9"             
#> [184] "AB(R)"             "FBl2"              "VMNP"             
#> [187] "PB(R3)"            "AL-VM7d(R)"        "PENP"             
#> [190] "GOR(L)"            "AL-DC2(R)"         "AL-VA3(R)"        
#> [193] "lACA(R)"           "a'2(R)"            "NO3(L)"           
#> [196] "b1(R)"             "PB(L4)"            "AL-VA5(R)"        
#> [199] "SCL(R)"            "EBr6"              "PB(L1)"           
#> [202] "AL-DM5(L)"         "g3(R)"             "FBl6"             
#> [205] "AL-VC4(R)"         "GOR(R)"            "mALT(L)"          
#> [208] "AL-DL1(R)"         "b2(R)"             "BU(R)"            
#> [211] "VES(R)"            "AL-DA4m(R)"        "EBr3d"            
#> [214] "MB(+ACA)(R)"       "ATL(R)"            "AL-DA4l(R)"       
#> [217] "AL-DP1l(R)"        "AL-VA6(R)"         "AL-DM5(R)"        
#> [220] "hemibrain"         "PED(R)"            "EPA(R)"           
#> [223] "AL-DA2(R)"         "OL(R)"             "AL-VC3m(R)"       
#> [226] "AL-VP4(R)"         "AL-DC4(L)"         "AL-DM1(L)"        
#> [229] "EBr3pw"            "bL(R)"            
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method   path                                           
#> [1,] "GET"    "/api/custom/custom"                           
#> [2,] "DELETE" "/api"                                         
#> [3,] "POST"   "/api/v:ver/npexplorer/celltype/:dataset/:type"
#> [4,] "GET"    "/oauth2callback"                              
#> [5,] "GET"    "/api/v:ver/dbmeta/database"                   
#> [6,] "GET"    "/api/dbmeta/instances"                        
#>      name                                                                                 
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"      
#> [2,] "github.com/labstack/echo.glob..func1"                                               
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                 
#> [4,] "github.com/janelia-flyem/echo-secure.oauthCallbackHandler"                          
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                 
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDataInstances-fm"
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }