Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "SMP(L)"            "PB(R7)"            "ME(R)"            
#>   [4] "PB(L6)"            "CA(R)"             "a3(R)"            
#>   [7] "PB(L1)"            "gL(L)"             "gL(R)"            
#>  [10] "ATL(R)"            "EPA(L)"            "SMP(R)"           
#>  [13] "IB"                "SCL(L)"            "RUB(R)"           
#>  [16] "AOTU(R)"           "PB(R8)"            "ROB(R)"           
#>  [19] "a'1(R)"            "SNP(L)"            "GNG"              
#>  [22] "b'L(L)"            "VMNP"              "MB(+ACA)(R)"      
#>  [25] "VES(L)"            "AOT(R)"            "vACA(R)"          
#>  [28] "g1(R)"             "EBr1"              "SIP(R)"           
#>  [31] "NO(R)"             "mALT(R)"           "POC"              
#>  [34] "GA(R)"             "NO3(R)"            "SNP(R)"           
#>  [37] "EBr2r4"            "FBl4"              "LX(L)"            
#>  [40] "b'2(R)"            "lACA(R)"           "CRE(R)"           
#>  [43] "AL(L)"             "LO(R)"             "PB(R3)"           
#>  [46] "NO2(R)"            "LAL(L)"            "LAL(-GA)(R)"      
#>  [49] "FBl5"              "LH(R)"             "NO(L)"            
#>  [52] "a'3(R)"            "NO1(R)"            "FBl2"             
#>  [55] "SCL(R)"            "SLP(R)"            "WED(R)"           
#>  [58] "PB(L8)"            "EPA(R)"            "AVLP(R)"          
#>  [61] "OL(R)"             "ATL(L)"            "PB(L3)"           
#>  [64] "SIP(L)"            "PED(R)"            "EBr3d"            
#>  [67] "PVLP(R)"           "AME(R)"            "LOP(R)"           
#>  [70] "NO"                "FBl7"              "CA(L)"            
#>  [73] "EBr5"              "ICL(L)"            "FBl3"             
#>  [76] "ICL(R)"            "AB(L)"             "PB(L5)"           
#>  [79] "a1(R)"             "SAD"               "g2(R)"            
#>  [82] "EBr6"              "SPS(L)"            "PB(R2)"           
#>  [85] "AL(R)"             "FB"                "CRE(L)"           
#>  [88] "FBl1"              "CAN(R)"            "LAL(R)"           
#>  [91] "b'1(R)"            "NO2(L)"            "a2(R)"            
#>  [94] "GOR(L)"            "MB(L)"             "b'L(R)"           
#>  [97] "FLA(R)"            "PB"                "a'2(R)"           
#> [100] "b1(R)"             "PB(L4)"            "VLNP(R)"          
#> [103] "PB(R5)"            "hemibrain"         "SPS(R)"           
#> [106] "aL(R)"             "MB(R)"             "PB(L2)"           
#> [109] "CRE(-ROB,-RUB)(R)" "BU(L)"             "PENP"             
#> [112] "NO1(L)"            "aL(L)"             "INP"              
#> [115] "PB(R6)"            "LX(R)"             "PB(L7)"           
#> [118] "SAD(-AMMC)"        "GF(R)"             "GC"               
#> [121] "g4(R)"             "EBr3am"            "PLP(R)"           
#> [124] "a'L(R)"            "PB(R9)"            "VES(R)"           
#> [127] "b2(R)"             "PB(L9)"            "bL(L)"            
#> [130] "AMMC"              "PRW"               "EBr3pw"           
#> [133] "RUB(L)"            "PB(R4)"            "bL(R)"            
#> [136] "dACA(R)"           "EB"                "FBl6"             
#> [139] "FBl9"              "IPS(R)"            "FBl8"             
#> [142] "a'L(L)"            "CX"                "PB(R1)"           
#> [145] "GOR(R)"            "mALT(L)"           "g3(R)"            
#> [148] "BU(R)"             "NO3(L)"            "g5(R)"            
#> [151] "AB(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "SPS(R)"            "INP"               "AOT(R)"           
#>   [4] "PRW"               "SCL(R)"            "PB"               
#>   [7] "FBl5"              "AL(L)"             "PVLP(R)"          
#>  [10] "FBl7"              "PB(L5)"            "NO"               
#>  [13] "ICL(L)"            "a'L(L)"            "AL-DL1(R)"        
#>  [16] "b'2(R)"            "dACA(R)"           "IPS(R)"           
#>  [19] "a3(R)"             "AL-DA4m(L)"        "AL-VA6(R)"        
#>  [22] "PB(L3)"            "gL(L)"             "b'L(L)"           
#>  [25] "VES(L)"            "FB-column3"        "OL(R)"            
#>  [28] "AL-VC1(R)"         "AL-VP5(R)"         "GOR(R)"           
#>  [31] "AL(R)"             "AL-DM1(R)"         "AL-VP4(R)"        
#>  [34] "AL-DM3(R)"         "AL-VM2(R)"         "MB(L)"            
#>  [37] "EBr2r4"            "PB(R3)"            "aL(L)"            
#>  [40] "LAL(L)"            "FBl1"              "bL(R)"            
#>  [43] "IB"                "PB(R6)"            "EPA(R)"           
#>  [46] "VES(R)"            "AL-DM6(R)"         "FB"               
#>  [49] "a'1(R)"            "AL-DP1l(R)"        "AL-DL2d(R)"       
#>  [52] "AL-VM1(R)"         "NO3(L)"            "AL-DM2(R)"        
#>  [55] "PB(R9)"            "NO(L)"             "PENP"             
#>  [58] "EPA(L)"            "AL-DA3(R)"         "AL-VC3m(R)"       
#>  [61] "ME(R)"             "SPS(L)"            "LX(L)"            
#>  [64] "AOTU(R)"           "GA(R)"             "FBl3"             
#>  [67] "AL-VP1m(R)"        "AL-VC3l(R)"        "AL-VM4(R)"        
#>  [70] "PB(L2)"            "EBr3am"            "mALT(R)"          
#>  [73] "AL-VA6(L)"         "AL-DA4l(R)"        "AL-VA2(R)"        
#>  [76] "AL-DL2v(R)"        "AL-V(R)"           "AL-VL1(R)"        
#>  [79] "AL-VM5v(R)"        "SCL(L)"            "SAD"              
#>  [82] "PED(R)"            "CA(R)"             "AB(R)"            
#>  [85] "AL-DP1m(R)"        "AB(L)"             "AL-D(L)"          
#>  [88] "AL-DM2(L)"         "SNP(L)"            "LH(R)"            
#>  [91] "a2(R)"             "PB(R4)"            "EBr3d"            
#>  [94] "EB"                "AMMC"              "AL-DC4(L)"        
#>  [97] "gL(R)"             "AL-VA1v(R)"        "AL-DM6(L)"        
#> [100] "CX"                "NO3(R)"            "g4(R)"            
#> [103] "AL-DL4(R)"         "AL-VA3(R)"         "PB(L6)"           
#> [106] "AL-VL2p(R)"        "vACA(R)"           "AL-DL4(L)"        
#> [109] "AL-DM5(R)"         "AL-DA2(L)"         "AL-DM1(L)"        
#> [112] "LOP(R)"            "NO2(L)"            "lACA(R)"          
#> [115] "g1(R)"             "b1(R)"             "AL-DL3(R)"        
#> [118] "AL-DA4m(R)"        "b'1(R)"            "LAL(-GA)(R)"      
#> [121] "PB(R8)"            "PB(R2)"            "AL-DA2(R)"        
#> [124] "AL-VP2(R)"         "aL(R)"             "g2(R)"            
#> [127] "g5(R)"             "AL-VC5(R)"         "WED(R)"           
#> [130] "FBl8"              "mALT(L)"           "CRE(R)"           
#> [133] "ROB(R)"            "NO2(R)"            "AL-VC4(R)"        
#> [136] "LX(R)"             "LO(R)"             "MB(R)"            
#> [139] "AL-VM7d(R)"        "CRE(-ROB,-RUB)(R)" "AL-DA3(L)"        
#> [142] "AL-VM7v(L)"        "NO(R)"             "CAN(R)"           
#> [145] "PLP(R)"            "AL-DC1(R)"         "SMP(L)"           
#> [148] "FBl2"              "AVLP(R)"           "RUB(R)"           
#> [151] "AL-VA7m(R)"        "NO1(R)"            "AL-DP1m(L)"       
#> [154] "EBr3pw"            "AL-DM4(R)"         "SIP(L)"           
#> [157] "AL-DL5(L)"         "CA(L)"             "GOR(L)"           
#> [160] "GC"                "PB(L7)"            "PB(R5)"           
#> [163] "g3(R)"             "PB(L1)"            "PB(L4)"           
#> [166] "AL-VP3(R)"         "BU(L)"             "POC"              
#> [169] "SNP(R)"            "AL-DM4(L)"         "FBl6"             
#> [172] "AL-DC4(R)"         "AL-D(R)"           "VMNP"             
#> [175] "AL-DC1(L)"         "PB(L9)"            "RUB(L)"           
#> [178] "EBr1"              "a'L(R)"            "CRE(-RUB)(L)"     
#> [181] "SMP(R)"            "PB(R7)"            "FBl4"             
#> [184] "AL-DC3(R)"         "AL-VA5(R)"         "GF(R)"            
#> [187] "ATL(R)"            "AME(R)"            "a'3(R)"           
#> [190] "AL-VP1d(R)"        "AL-DM3(L)"         "PB(R1)"           
#> [193] "bL(L)"             "AL-VC2(R)"         "AL-VA1d(R)"       
#> [196] "AL-VA4(R)"         "b'L(R)"            "AL-DM5(L)"        
#> [199] "AL-VM7d(L)"        "AL-DA1(R)"         "AL-VA7l(R)"       
#> [202] "EBr5"              "SIP(R)"            "MB(+ACA)(R)"      
#> [205] "GNG"               "b2(R)"             "ATL(L)"           
#> [208] "hemibrain"         "EBr6"              "LAL(R)"           
#> [211] "AL-VM5d(R)"        "a1(R)"             "AL-DC2(R)"        
#> [214] "AL-VL2a(R)"        "AL-VM3(R)"         "AL-VP1l(R)"       
#> [217] "NO1(L)"            "CRE(L)"            "FBl9"             
#> [220] "SLP(R)"            "VLNP(R)"           "ICL(R)"           
#> [223] "FLA(R)"            "a'2(R)"            "BU(R)"            
#> [226] "SAD(-AMMC)"        "PB(L8)"            "AL-DC2(L)"        
#> [229] "AL-DL5(R)"         "AL-VM7v(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.2`
#> $`hemibrain:v1.2`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2`$ROIs
#>   [1] "FBl1"              "PB(L4)"            "AL-DP1l(R)"       
#>   [4] "AL-VA1d(R)"        "AMMC"              "EBr1"             
#>   [7] "AL-DC2(L)"         "CA(L)"             "BU(R)"            
#>  [10] "AL-DA3(L)"         "AL-DM4(L)"         "SLP(R)"           
#>  [13] "AL-DA4l(R)"        "AOT(R)"            "PB(R1)"           
#>  [16] "IPS(R)"            "VES(L)"            "AL-DM1(R)"        
#>  [19] "AL-VA2(R)"         "EBr2r4"            "b'L(L)"           
#>  [22] "AL-VP2(R)"         "a'3(R)"            "AL-VM7d(L)"       
#>  [25] "a'1(R)"            "AL-DM2(R)"         "AL-VC3m(R)"       
#>  [28] "AL-VL1(R)"         "lACA(R)"           "b'L(R)"           
#>  [31] "PB(L8)"            "PB(R4)"            "VLNP(R)"          
#>  [34] "AL-DM6(L)"         "a'L(L)"            "AL-VA5(R)"        
#>  [37] "AL-VA6(R)"         "AL-VM5d(R)"        "LOP(R)"           
#>  [40] "aL(R)"             "g2(R)"             "EPA(R)"           
#>  [43] "NO2(L)"            "b'2(R)"            "LAL(L)"           
#>  [46] "a'L(R)"            "AL-DC4(L)"         "AL-DM1(L)"        
#>  [49] "RUB(R)"            "ATL(L)"            "SIP(L)"           
#>  [52] "AL-DA2(L)"         "CA(R)"             "AL-DM6(R)"        
#>  [55] "a2(R)"             "PB(L1)"            "AL-DA4m(L)"       
#>  [58] "AL-DC1(L)"         "a3(R)"             "OL(R)"            
#>  [61] "AL-DA1(R)"         "NO(L)"             "AL-DC1(R)"        
#>  [64] "AL-VC4(R)"         "CRE(R)"            "AL-VA1v(R)"       
#>  [67] "POC"               "FBl7"              "ATL(R)"           
#>  [70] "ICL(R)"            "MB(R)"             "AL-DP1m(R)"       
#>  [73] "PB(R6)"            "GF(R)"             "FBl3"             
#>  [76] "LH(R)"             "AL-DM5(R)"         "ICL(L)"           
#>  [79] "PB(R5)"            "GC"                "AL-VA6(L)"        
#>  [82] "PVLP(R)"           "MB(+ACA)(R)"       "FBl4"             
#>  [85] "CAN(R)"            "AL-VA4(R)"         "gL(L)"            
#>  [88] "CRE(L)"            "FBl8"              "LAL(R)"           
#>  [91] "PB(L5)"            "AL-VM4(R)"         "EBr3am"           
#>  [94] "PB(R9)"            "AL-DC2(R)"         "MB(L)"            
#>  [97] "EBr6"              "NO1(R)"            "PB"               
#> [100] "AOTU(R)"           "g4(R)"             "AL-DM5(L)"        
#> [103] "CRE(-RUB)(L)"      "PB(R2)"            "NO3(R)"           
#> [106] "EBr3d"             "EBr3pw"            "g3(R)"            
#> [109] "WED(R)"            "BU(L)"             "AL-VM7v(R)"       
#> [112] "AL-DL4(L)"         "SAD(-AMMC)"        "PB(R3)"           
#> [115] "PB(L2)"            "PB(R7)"            "bL(L)"            
#> [118] "AL-DM2(L)"         "AL-VP5(R)"         "GOR(L)"           
#> [121] "NO1(L)"            "SNP(R)"            "PB(L6)"           
#> [124] "GNG"               "AL-VM7d(R)"        "vACA(R)"          
#> [127] "NO"                "gL(R)"             "PB(L9)"           
#> [130] "NO3(L)"            "AL-DC4(R)"         "AL-DL4(R)"        
#> [133] "AL-VC5(R)"         "AL-VP1d(R)"        "FLA(R)"           
#> [136] "AL-DM3(L)"         "LO(R)"             "FBl2"             
#> [139] "AL-VC3l(R)"        "CRE(-ROB,-RUB)(R)" "CX"               
#> [142] "AL-DL2v(R)"        "AL-DL3(R)"         "PED(R)"           
#> [145] "ROB(R)"            "RUB(L)"            "b'1(R)"           
#> [148] "LX(R)"             "FBl6"              "AL-VC1(R)"        
#> [151] "SCL(R)"            "a'2(R)"            "FB-column3"       
#> [154] "AL-DA3(R)"         "PENP"              "AL(L)"            
#> [157] "FBl9"              "AB(R)"             "AL-VP1m(R)"       
#> [160] "b2(R)"             "SNP(L)"            "AL-DL1(R)"        
#> [163] "AL-DL5(R)"         "AL-VP1l(R)"        "SMP(R)"           
#> [166] "NO(R)"             "PB(L7)"            "PB(R8)"           
#> [169] "AL-D(L)"           "FBl5"              "FB"               
#> [172] "NO2(R)"            "AL(R)"             "AL-DA2(R)"        
#> [175] "AB(L)"             "b1(R)"             "bL(R)"            
#> [178] "LAL(-GA)(R)"       "SPS(R)"            "SIP(R)"           
#> [181] "AL-DM4(R)"         "AL-VL2p(R)"        "mALT(R)"          
#> [184] "AL-VC2(R)"         "AL-VM2(R)"         "EBr5"             
#> [187] "dACA(R)"           "EB"                "PLP(R)"           
#> [190] "AL-VM5v(R)"        "AL-DL5(L)"         "AME(R)"           
#> [193] "GA(R)"             "AL-DL2d(R)"        "AL-D(R)"          
#> [196] "AL-VP3(R)"         "IB"                "mALT(L)"          
#> [199] "ME(R)"             "GOR(R)"            "AL-VA7l(R)"       
#> [202] "SMP(L)"            "AL-VA3(R)"         "INP"              
#> [205] "AL-VM3(R)"         "SAD"               "AL-VA7m(R)"       
#> [208] "PRW"               "VMNP"              "g5(R)"            
#> [211] "AVLP(R)"           "LX(L)"             "AL-DM3(R)"        
#> [214] "AL-VL2a(R)"        "hemibrain"         "AL-DP1m(L)"       
#> [217] "aL(L)"             "AL-DA4m(R)"        "AL-V(R)"          
#> [220] "PB(L3)"            "g1(R)"             "SPS(L)"           
#> [223] "VES(R)"            "SCL(L)"            "AL-VM7v(L)"       
#> [226] "AL-VP4(R)"         "a1(R)"             "EPA(L)"           
#> [229] "AL-DC3(R)"         "AL-VM1(R)"        
#> 
#> $`hemibrain:v1.2`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method  path                                     
#> [1,] "GET"   "/api"                                   
#> [2,] "GET"   "/api/v:ver/npexplorer/simpleconnections"
#> [3,] "POST"  "/api/npexplorer/roisinneuron"           
#> [4,] "GET"   "/api/npexplorer/autapses"               
#> [5,] "POST"  "/api/raw/cypher/cypher"                 
#> [6,] "TRACE" "/api"                                   
#>      name                                                                                        
#> [1,] "github.com/labstack/echo.glob..func1"                                                      
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"    
#> [5,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [6,] "github.com/labstack/echo.glob..func1"                                                      

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }