Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "EBr3pw"            "AL-VA5(R)"         "b2(R)"            
#>   [4] "AL-DM2(L)"         "aL(L)"             "AL-VC5(R)"        
#>   [7] "PVLP(R)"           "FBl4"              "AL-VC4(R)"        
#>  [10] "AL-VM7d(R)"        "AB(L)"             "PB(R6)"           
#>  [13] "SAD"               "FBl9"              "a2(R)"            
#>  [16] "AL-DM5(L)"         "PB(L2)"            "AL-DL2d(R)"       
#>  [19] "PB(R5)"            "GF(R)"             "AL-VP5(R)"        
#>  [22] "PB(L5)"            "AVLP(R)"           "AL-DL1(R)"        
#>  [25] "LOP(R)"            "FLA(R)"            "AL-DM3(L)"        
#>  [28] "GOR(R)"            "MB(R)"             "NO1(R)"           
#>  [31] "FBl6"              "NO(L)"             "SCL(L)"           
#>  [34] "AL-VM3(R)"         "SMP(R)"            "b'L(R)"           
#>  [37] "CAN(R)"            "GNG"               "bL(R)"            
#>  [40] "PB(L9)"            "ICL(L)"            "a'3(R)"           
#>  [43] "a1(R)"             "FBl7"              "AL-VC1(R)"        
#>  [46] "ME(R)"             "PB(R1)"            "AL-DL2v(R)"       
#>  [49] "AL-DL3(R)"         "AOT(R)"            "EBr3am"           
#>  [52] "b'2(R)"            "GC"                "AL-DM3(R)"        
#>  [55] "PB(R8)"            "AL-DA2(L)"         "PB"               
#>  [58] "AL-DC2(L)"         "CRE(-RUB)(L)"      "AL(R)"            
#>  [61] "PB(L3)"            "mALT(R)"           "AL(L)"            
#>  [64] "LH(R)"             "FB"                "PLP(R)"           
#>  [67] "SCL(R)"            "AL-DL4(L)"         "b'L(L)"           
#>  [70] "AL-DM5(R)"         "AL-VM7v(L)"        "bL(L)"            
#>  [73] "aL(R)"             "PB(L1)"            "PB(L6)"           
#>  [76] "a'1(R)"            "PB(L7)"            "SLP(R)"           
#>  [79] "hemibrain"         "GOR(L)"            "AL-DA4m(R)"       
#>  [82] "AL-VP1m(R)"        "NO2(R)"            "PB(R4)"           
#>  [85] "AL-VC3l(R)"        "EBr5"              "EPA(R)"           
#>  [88] "PED(R)"            "CA(L)"             "SMP(L)"           
#>  [91] "AL-VM7v(R)"        "g5(R)"             "PRW"              
#>  [94] "EBr2r4"            "INP"               "AL-VC2(R)"        
#>  [97] "SIP(L)"            "CA(R)"             "AL-VA1v(R)"       
#> [100] "IB112"             "AL-DA4l(R)"        "FBl5"             
#> [103] "AL-VC3m(R)"        "GA(R)"             "AL-VA6(L)"        
#> [106] "FBl1"              "AME(R)"            "LAL(R)"           
#> [109] "CRE(-ROB,-RUB)(R)" "SPS(R)"            "AL-DC1(L)"        
#> [112] "AL-VM1(R)"         "a'2(R)"            "SNP(L)"           
#> [115] "AL-DA3(R)"         "VES(L)"            "AL-VA7m(R)"       
#> [118] "LX(L)"             "CX"                "ROB(R)"           
#> [121] "CRE(L)"            "gL(L)"             "AL-VP4(R)"        
#> [124] "NO1(L)"            "EB"                "FBl8"             
#> [127] "AL-DM1(R)"         "AL-DA4m(L)"        "AL-DM6(L)"        
#> [130] "a3(R)"             "AL-D(R)"           "AL-VP2(R)"        
#> [133] "VLNP(R)"           "g2(R)"             "PB(R3)"           
#> [136] "NO3(L)"            "VES(R)"            "AL-D(L)"          
#> [139] "EPA(L)"            "LAL(-GA)(R)"       "CRE(R)"           
#> [142] "dACA(R)"           "AL-DP1l(R)"        "IPS(R)"           
#> [145] "AL-DA2(R)"         "AL-DM4(R)"         "AL-V(R)"          
#> [148] "IB"                "MB(L)"             "EBr3d"            
#> [151] "LAL(L)"            "FBl2"              "AL-VA1d(R)"       
#> [154] "AL-VL1(R)"         "AL-VM5v(R)"        "EBr1"             
#> [157] "AL-DC4(R)"         "PB(L4)"            "AL-VP1d(R)"       
#> [160] "gL(R)"             "AL-DL5(R)"         "AL-VA4(R)"        
#> [163] "g4(R)"             "OL(R)"             "PB(R7)"           
#> [166] "AL-DC4(L)"         "AL-DP1m(L)"        "b1(R)"            
#> [169] "AL-VA2(R)"         "g3(R)"             "SIP(R)"           
#> [172] "AL-DC3(R)"         "AL-VL2a(R)"        "AL-VM2(R)"        
#> [175] "AL-VM5d(R)"        "POC"               "mALT(L)"          
#> [178] "NO"                "LO(R)"             "SAD(-AMMC)"       
#> [181] "FBl3"              "a'L(L)"            "b'1(R)"           
#> [184] "AL-DM1(L)"         "PB(R9)"            "AL-VA6(R)"        
#> [187] "AL-VL2p(R)"        "NO2(L)"            "AOTU(R)"          
#> [190] "AL-DL4(R)"         "AL-DM6(R)"         "AL-VM4(R)"        
#> [193] "RUB(R)"            "EBr6"              "AL-DC1(R)"        
#> [196] "AL-DM2(R)"         "lACA(R)"           "AL-DA1(R)"        
#> [199] "AL-VM7d(L)"        "MB(+ACA)(R)"       "ATL(R)"           
#> [202] "g1(R)"             "BU(L)"             "NO(R)"            
#> [205] "ICL(R)"            "RUB(L)"            "AL-VP3(R)"        
#> [208] "vACA(R)"           "ATL(L)"            "LX(R)"            
#> [211] "a'L(R)"            "AL-VA7l(R)"        "AMMC"             
#> [214] "PB(R2)"            "AL-VA3(R)"         "FB-column3"       
#> [217] "AL-DA3(L)"         "AL-DL5(L)"         "AL-DM4(L)"        
#> [220] "AL-DC2(R)"         "SNP(R)"            "BU(R)"            
#> [223] "PENP"              "AL-VP1l(R)"        "PB(L8)"           
#> [226] "VMNP"              "AL-DP1m(R)"        "WED(R)"           
#> [229] "NO3(R)"            "SPS(L)"            "AB(R)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "mALT(L)"           "NO3(R)"            "dACA(R)"          
#>   [4] "ME(R)"             "PB(L5)"            "PB"               
#>   [7] "PLP(R)"            "POC"               "SIP(R)"           
#>  [10] "FBl5"              "EPA(L)"            "VES(L)"           
#>  [13] "VMNP"              "ICL(R)"            "CA(L)"            
#>  [16] "FBl4"              "FBl8"              "PB(L2)"           
#>  [19] "PB(R7)"            "SNP(R)"            "AL(L)"            
#>  [22] "CRE(L)"            "INP"               "NO2(L)"           
#>  [25] "EBr6"              "LX(R)"             "AB(R)"            
#>  [28] "CAN(R)"            "a3(R)"             "vACA(R)"          
#>  [31] "SPS(R)"            "PB(R8)"            "SAD(-AMMC)"       
#>  [34] "FBl3"              "hemibrain"         "PED(R)"           
#>  [37] "FBl7"              "PB(R5)"            "AOTU(R)"          
#>  [40] "SCL(R)"            "a'1(R)"            "LX(L)"            
#>  [43] "EB"                "SMP(L)"            "GC"               
#>  [46] "g2(R)"             "BU(L)"             "b'L(L)"           
#>  [49] "AOT(R)"            "SCL(L)"            "g1(R)"            
#>  [52] "b2(R)"             "b'L(R)"            "PB(R4)"           
#>  [55] "PB(L1)"            "PB(L8)"            "OL(R)"            
#>  [58] "EBr3am"            "WED(R)"            "CRE(R)"           
#>  [61] "EBr1"              "CA(R)"             "ROB(R)"           
#>  [64] "NO(R)"             "bL(R)"             "MB(R)"            
#>  [67] "b'2(R)"            "GA(R)"             "GOR(L)"           
#>  [70] "g4(R)"             "FLA(R)"            "a'2(R)"           
#>  [73] "GOR(R)"            "NO"                "gL(R)"            
#>  [76] "b1(R)"             "LAL(R)"            "LAL(L)"           
#>  [79] "lACA(R)"           "mALT(R)"           "BU(R)"            
#>  [82] "VES(R)"            "AMMC"              "MB(L)"            
#>  [85] "LH(R)"             "RUB(L)"            "AB(L)"            
#>  [88] "PB(L3)"            "SNP(L)"            "NO2(R)"           
#>  [91] "EPA(R)"            "PB(R1)"            "PB(R3)"           
#>  [94] "GNG"               "a'L(L)"            "a1(R)"            
#>  [97] "PB(R9)"            "MB(+ACA)(R)"       "NO(L)"            
#> [100] "RUB(R)"            "PENP"              "SAD"              
#> [103] "PRW"               "EBr3d"             "CRE(-ROB,-RUB)(R)"
#> [106] "EBr5"              "PVLP(R)"           "ATL(L)"           
#> [109] "SIP(L)"            "LO(R)"             "PB(L4)"           
#> [112] "IB"                "GF(R)"             "g3(R)"            
#> [115] "b'1(R)"            "LOP(R)"            "VLNP(R)"          
#> [118] "a'3(R)"            "LAL(-GA)(R)"       "gL(L)"            
#> [121] "g5(R)"             "AL(R)"             "bL(L)"            
#> [124] "FBl9"              "AVLP(R)"           "IPS(R)"           
#> [127] "a2(R)"             "NO1(R)"            "EBr2r4"           
#> [130] "aL(L)"             "FBl2"              "AME(R)"           
#> [133] "SLP(R)"            "SMP(R)"            "PB(L9)"           
#> [136] "NO3(L)"            "EBr3pw"            "ICL(L)"           
#> [139] "FBl1"              "FBl6"              "aL(R)"            
#> [142] "FB"                "a'L(R)"            "PB(L7)"           
#> [145] "PB(R6)"            "NO1(L)"            "PB(R2)"           
#> [148] "ATL(R)"            "SPS(L)"            "PB(L6)"           
#> [151] "CX"               
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DA3(L)"         "AL-DM3(L)"         "AL-VA6(L)"        
#>   [4] "LH(R)"             "IPS(R)"            "AL-VL1(R)"        
#>   [7] "PB(R6)"            "g1(R)"             "PB(R2)"           
#>  [10] "AL-VA7m(R)"        "AL-VC3m(R)"        "AL-DA1(R)"        
#>  [13] "CAN(R)"            "INP"               "PB(L5)"           
#>  [16] "LAL(-GA)(R)"       "PB(L9)"            "AL-DC1(R)"        
#>  [19] "AL-DC4(R)"         "lACA(R)"           "AL(L)"            
#>  [22] "PB(R9)"            "SIP(L)"            "a3(R)"            
#>  [25] "PB(L2)"            "FBl1"              "AL-VA2(R)"        
#>  [28] "AL-VA5(R)"         "AL-VA7l(R)"        "BU(R)"            
#>  [31] "CA(R)"             "AL-DL2v(R)"        "CRE(R)"           
#>  [34] "PB"                "PB(L3)"            "LX(L)"            
#>  [37] "NO3(L)"            "FBl5"              "PB(R3)"           
#>  [40] "AL-DC3(R)"         "AL-DC1(L)"         "LX(R)"            
#>  [43] "gL(L)"             "AL-VM2(R)"         "GA(R)"            
#>  [46] "EPA(R)"            "AL-DA4m(L)"        "AL-DC4(L)"        
#>  [49] "EBr1"              "AL(R)"             "hemibrain"        
#>  [52] "g2(R)"             "AL-DA4l(R)"        "EBr3d"            
#>  [55] "FBl7"              "bL(R)"             "a'3(R)"           
#>  [58] "b'2(R)"            "a'1(R)"            "ICL(R)"           
#>  [61] "CA(L)"             "NO1(L)"            "b'L(R)"           
#>  [64] "AL-DA2(L)"         "CRE(-RUB)(L)"      "AL-VC1(R)"        
#>  [67] "LAL(L)"            "PRW"               "AL-DP1m(L)"       
#>  [70] "ICL(L)"            "AMMC"              "PB(R7)"           
#>  [73] "AL-D(L)"           "AL-DM2(L)"         "AL-DL4(R)"        
#>  [76] "PB(R5)"            "NO1(R)"            "AL-DA2(R)"        
#>  [79] "AL-DP1m(R)"        "WED(R)"            "AL-VP4(R)"        
#>  [82] "AL-VP2(R)"         "AL-VA1d(R)"        "a1(R)"            
#>  [85] "GF(R)"             "PB(L1)"            "a'L(R)"           
#>  [88] "AL-VM7v(L)"        "AL-DC2(R)"         "PENP"             
#>  [91] "AL-DC2(L)"         "NO(R)"             "aL(R)"            
#>  [94] "AL-DM5(L)"         "CRE(L)"            "AL-DM1(R)"        
#>  [97] "VES(R)"            "POC"               "mALT(L)"          
#> [100] "AL-V(R)"           "SCL(R)"            "PB(R8)"           
#> [103] "SIP(R)"            "PB(L7)"            "GNG"              
#> [106] "MB(+ACA)(R)"       "AL-VA6(R)"         "MB(L)"            
#> [109] "b1(R)"             "AL-DL1(R)"         "ATL(L)"           
#> [112] "PB(L8)"            "PLP(R)"            "a'2(R)"           
#> [115] "AL-DP1l(R)"        "BU(L)"             "gL(R)"            
#> [118] "g5(R)"             "PED(R)"            "SNP(L)"           
#> [121] "FLA(R)"            "VES(L)"            "AL-DA4m(R)"       
#> [124] "AL-DL2d(R)"        "AL-VL2a(R)"        "SMP(L)"           
#> [127] "AOTU(R)"           "NO3(R)"            "FBl8"             
#> [130] "AME(R)"            "FB"                "AL-VM7d(L)"       
#> [133] "NO"                "FB-column3"        "AL-VM7d(R)"       
#> [136] "mALT(R)"           "VMNP"              "AB(L)"            
#> [139] "GOR(R)"            "AL-DM6(R)"         "AL-VM4(R)"        
#> [142] "AL-VP1l(R)"        "SCL(L)"            "AL-DL4(L)"        
#> [145] "AL-VA1v(R)"        "AL-VL2p(R)"        "AL-VM5d(R)"       
#> [148] "FBl2"              "a'L(L)"            "CRE(-ROB,-RUB)(R)"
#> [151] "AL-DL3(R)"         "AL-DL5(R)"         "AL-DM2(R)"        
#> [154] "AL-VC4(R)"         "vACA(R)"           "EBr5"             
#> [157] "EBr3pw"            "NO(L)"             "RUB(R)"           
#> [160] "CX"                "SAD"               "NO2(R)"           
#> [163] "PB(L4)"            "AL-DA3(R)"         "IB"               
#> [166] "AL-DM6(L)"         "PB(L6)"            "bL(L)"            
#> [169] "OL(R)"             "PB(R1)"            "SAD(-AMMC)"       
#> [172] "AVLP(R)"           "AL-VC5(R)"         "VLNP(R)"          
#> [175] "EBr6"              "LO(R)"             "LAL(R)"           
#> [178] "SLP(R)"            "AL-VM3(R)"         "FBl9"             
#> [181] "AB(R)"             "MB(R)"             "AOT(R)"           
#> [184] "dACA(R)"           "RUB(L)"            "AL-DM5(R)"        
#> [187] "AL-VC2(R)"         "NO2(L)"            "PVLP(R)"          
#> [190] "AL-DM4(R)"         "AL-VA3(R)"         "AL-VA4(R)"        
#> [193] "b'1(R)"            "FBl4"              "b'L(L)"           
#> [196] "AL-DM3(R)"         "SMP(R)"            "GC"               
#> [199] "EB"                "FBl6"              "AL-VM7v(R)"       
#> [202] "SNP(R)"            "AL-VP3(R)"         "AL-VP1d(R)"       
#> [205] "GOR(L)"            "EBr2r4"            "ATL(R)"           
#> [208] "AL-VM1(R)"         "ME(R)"             "g4(R)"            
#> [211] "FBl3"              "AL-VP5(R)"         "a2(R)"            
#> [214] "ROB(R)"            "AL-DM1(L)"         "AL-DM4(L)"        
#> [217] "PB(R4)"            "LOP(R)"            "SPS(R)"           
#> [220] "AL-DL5(L)"         "aL(L)"             "b2(R)"            
#> [223] "AL-D(R)"           "AL-VC3l(R)"        "SPS(L)"           
#> [226] "EBr3am"            "g3(R)"             "EPA(L)"           
#> [229] "AL-VM5v(R)"        "AL-VP1m(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "SMP(L)"            "g2(R)"             "MB(+ACA)(R)"      
#>   [4] "PB(R4)"            "AL-VA7l(R)"        "PB(R6)"           
#>   [7] "AL-DM3(L)"         "RUB(L)"            "b'L(L)"           
#>  [10] "AL-VM7v(R)"        "SNP(R)"            "PB(L9)"           
#>  [13] "IB"                "NO3(R)"            "AL-DM5(L)"        
#>  [16] "AL-DL5(L)"         "PB(R1)"            "NO(R)"            
#>  [19] "SLP(R)"            "AL-DL1(R)"         "EBr1"             
#>  [22] "b1(R)"             "AL-VA4(R)"         "LOP(R)"           
#>  [25] "ME(R)"             "EBr2r4"            "AL-VM7d(L)"       
#>  [28] "AL-DL3(R)"         "SMP(R)"            "AL-VC1(R)"        
#>  [31] "SAD"               "SCL(R)"            "FLA(R)"           
#>  [34] "PB(R9)"            "AL-DC2(R)"         "AL-VA6(R)"        
#>  [37] "g3(R)"             "PRW"               "g1(R)"            
#>  [40] "LX(R)"             "VLNP(R)"           "a'3(R)"           
#>  [43] "FBl4"              "AL(R)"             "b'1(R)"           
#>  [46] "EBr5"              "NO1(R)"            "a'L(R)"           
#>  [49] "FBl5"              "AL-VM5d(R)"        "ICL(L)"           
#>  [52] "AL-VC2(R)"         "EBr3am"            "EPA(R)"           
#>  [55] "gL(R)"             "PVLP(R)"           "AL-VL1(R)"        
#>  [58] "AOTU(R)"           "AL-DM1(R)"         "AL-VA7m(R)"       
#>  [61] "ATL(L)"            "a2(R)"             "GC"               
#>  [64] "FBl7"              "AL-DP1m(R)"        "b2(R)"            
#>  [67] "AL-DM2(L)"         "IPS(R)"            "GNG"              
#>  [70] "a'L(L)"            "CAN(R)"            "CX"               
#>  [73] "AL-VA1d(R)"        "AL-VA3(R)"         "a1(R)"            
#>  [76] "PB(L1)"            "AL-DC1(L)"         "CRE(L)"           
#>  [79] "AL-DM3(R)"         "WED(R)"            "NO1(L)"           
#>  [82] "BU(R)"             "AL-DM1(L)"         "FBl2"             
#>  [85] "PB(R2)"            "EBr3d"             "PENP"             
#>  [88] "PB(L7)"            "LAL(R)"            "AL-DL5(R)"        
#>  [91] "b'2(R)"            "hemibrain"         "PB(R7)"           
#>  [94] "AL-DC4(R)"         "CA(L)"             "gL(L)"            
#>  [97] "AL-VP3(R)"         "a'2(R)"            "PB(R3)"           
#> [100] "AB(R)"             "bL(L)"             "AL-DC1(R)"        
#> [103] "lACA(R)"           "SAD(-AMMC)"        "AME(R)"           
#> [106] "b'L(R)"            "dACA(R)"           "AL-DA4m(L)"       
#> [109] "PB(L4)"            "SCL(L)"            "AL-DC2(L)"        
#> [112] "EBr6"              "AL-DL4(L)"         "NO"               
#> [115] "AL-DP1l(R)"        "PB(R5)"            "AL(L)"            
#> [118] "AL-DL4(R)"         "AL-VA2(R)"         "a'1(R)"           
#> [121] "FBl3"              "bL(R)"             "AL-VA1v(R)"       
#> [124] "AL-VC3m(R)"        "AL-VA6(L)"         "FBl8"             
#> [127] "INP"               "NO(L)"             "AL-DL2d(R)"       
#> [130] "AL-VC5(R)"         "ATL(R)"            "NO2(R)"           
#> [133] "AL-DM6(L)"         "LX(L)"             "AL-DM6(R)"        
#> [136] "aL(R)"             "AL-DL2v(R)"        "AL-DM4(R)"        
#> [139] "AL-VM2(R)"         "vACA(R)"           "AL-VC4(R)"        
#> [142] "PB(L3)"            "GA(R)"             "AL-DC4(L)"        
#> [145] "LO(R)"             "FB-column3"        "EPA(L)"           
#> [148] "SIP(L)"            "SPS(L)"            "mALT(R)"          
#> [151] "GOR(R)"            "AL-VL2a(R)"        "AL-VM7d(R)"       
#> [154] "SPS(R)"            "FBl6"              "CA(R)"            
#> [157] "OL(R)"             "AL-DA3(R)"         "AL-DM2(R)"        
#> [160] "SNP(L)"            "VES(R)"            "GF(R)"            
#> [163] "AVLP(R)"           "AL-VP2(R)"         "CRE(-ROB,-RUB)(R)"
#> [166] "FB"                "AL-VM4(R)"         "PB"               
#> [169] "g4(R)"             "SIP(R)"            "AL-VM5v(R)"       
#> [172] "AL-VP4(R)"         "NO2(L)"            "AL-DA3(L)"        
#> [175] "AL-V(R)"           "FBl1"              "EB"               
#> [178] "AL-VA5(R)"         "AL-VP1m(R)"        "AL-VP1l(R)"       
#> [181] "AOT(R)"            "MB(L)"             "NO3(L)"           
#> [184] "AL-DA2(R)"         "AL-DM5(R)"         "AL-DM4(L)"        
#> [187] "ROB(R)"            "g5(R)"             "ICL(R)"           
#> [190] "PB(L5)"            "GOR(L)"            "AL-VM3(R)"        
#> [193] "PLP(R)"            "a3(R)"             "AL-DC3(R)"        
#> [196] "PB(R8)"            "AL-DA4l(R)"        "AL-DA4m(R)"       
#> [199] "AL-D(R)"           "LH(R)"             "AL-DA2(L)"        
#> [202] "LAL(L)"            "PED(R)"            "CRE(R)"           
#> [205] "PB(L6)"            "VES(L)"            "AL-VC3l(R)"       
#> [208] "LAL(-GA)(R)"       "AL-D(L)"           "FBl9"             
#> [211] "MB(R)"             "AL-VM1(R)"         "AB(L)"            
#> [214] "BU(L)"             "AL-DA1(R)"         "PB(L2)"           
#> [217] "AMMC"              "mALT(L)"           "VMNP"             
#> [220] "AL-DP1m(L)"        "CRE(-RUB)(L)"      "aL(L)"            
#> [223] "AL-VL2p(R)"        "POC"               "PB(L8)"           
#> [226] "AL-VM7v(L)"        "EBr3pw"            "AL-VP5(R)"        
#> [229] "AL-VP1d(R)"        "RUB(R)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method   path                                     
#> [1,] "GET"    "/api/dbmeta/datasets"                   
#> [2,] "GET"    "/api/v:ver/npexplorer/completeness"     
#> [3,] "GET"    "/api/v:ver/cached/roicompleteness"      
#> [4,] "REPORT" "/api/*"                                 
#> [5,] "GET"    "/api/custom/custom"                     
#> [6,] "POST"   "/api/v:ver/roimeshes/mesh/:dataset/:roi"
#>      name                                                                            
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatasets-fm"
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"            
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"            
#> [4,] "github.com/labstack/echo/v4.glob..func1"                                       
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm" 
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"            
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }