Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "BU(R)"             "AL-DM4(L)"         "FBl6"             
#>   [4] "PB(R5)"            "AMMC"              "b'2(R)"           
#>   [7] "SPS(R)"            "BU(L)"             "AL-DM6(L)"        
#>  [10] "AL-VL1(R)"         "EBr5"              "ROB(R)"           
#>  [13] "AL(R)"             "AL-DP1m(R)"        "b'1(R)"           
#>  [16] "aL(R)"             "MB(+ACA)(R)"       "NO2(R)"           
#>  [19] "LH(R)"             "AL-VP1l(R)"        "AB(L)"            
#>  [22] "EBr3am"            "dACA(R)"           "OL(R)"            
#>  [25] "AL-VA2(R)"         "RUB(R)"            "MB(L)"            
#>  [28] "AL-DA3(R)"         "AL-DM4(R)"         "AL-VC2(R)"        
#>  [31] "AL-VA1d(R)"        "AL-VM7d(R)"        "POC"              
#>  [34] "mALT(R)"           "g1(R)"             "VMNP"             
#>  [37] "AL-DM1(L)"         "AL-DM3(R)"         "CRE(R)"           
#>  [40] "a'2(R)"            "AL-DA4m(L)"        "b'L(L)"           
#>  [43] "GOR(L)"            "FLA(R)"            "PB(L4)"           
#>  [46] "PB(R3)"            "PB(R8)"            "g4(R)"            
#>  [49] "vACA(R)"           "a1(R)"             "aL(L)"            
#>  [52] "VES(L)"            "AL-VC3m(R)"        "AL-VM2(R)"        
#>  [55] "AL-DC2(L)"         "AL-VM3(R)"         "CRE(L)"           
#>  [58] "AL-DM2(R)"         "AL-VP5(R)"         "LAL(L)"           
#>  [61] "CRE(-RUB)(L)"      "ATL(R)"            "INP"              
#>  [64] "MB(R)"             "LOP(R)"            "GF(R)"            
#>  [67] "LO(R)"             "AL-DA1(R)"         "NO1(L)"           
#>  [70] "AL-VM7d(L)"        "FBl3"              "SCL(L)"           
#>  [73] "SMP(L)"            "GA(R)"             "AL-DP1m(L)"       
#>  [76] "AL-DL2d(R)"        "AL-DM1(R)"         "PB(L5)"           
#>  [79] "PB(R4)"            "SIP(L)"            "EB"               
#>  [82] "bL(L)"             "AL-DL3(R)"         "AL-VP3(R)"        
#>  [85] "NO(R)"             "AL-VL2a(R)"        "AL-DC1(L)"        
#>  [88] "IPS(R)"            "GNG"               "GOR(R)"           
#>  [91] "AL-VM1(R)"         "gL(L)"             "NO1(R)"           
#>  [94] "EPA(R)"            "AL-DL5(L)"         "EPA(L)"           
#>  [97] "AL-VP2(R)"         "PB(L1)"            "IB112"            
#> [100] "AL-VM4(R)"         "PLP(R)"            "AL-DL4(L)"        
#> [103] "PVLP(R)"           "SNP(R)"            "AL-DM5(R)"        
#> [106] "a'3(R)"            "lACA(R)"           "b2(R)"            
#> [109] "SPS(L)"            "FBl1"              "AL-DL2v(R)"       
#> [112] "AL-DL4(R)"         "PED(R)"            "a'L(R)"           
#> [115] "FB-column3"        "AL-DL1(R)"         "AL-VM5d(R)"       
#> [118] "NO"                "g5(R)"             "a3(R)"            
#> [121] "IB"                "AL-VL2p(R)"        "AOTU(R)"          
#> [124] "VES(R)"            "AVLP(R)"           "AL-DA2(L)"        
#> [127] "AL-DM5(L)"         "AL-VM7v(L)"        "AL-DA4l(R)"       
#> [130] "AL-VC3l(R)"        "AL-VC1(R)"         "EBr1"             
#> [133] "PB(L8)"            "EBr2r4"            "PB(L9)"           
#> [136] "ICL(R)"            "AL-DA4m(R)"        "EBr6"             
#> [139] "AL-DA3(L)"         "CAN(R)"            "LAL(R)"           
#> [142] "AL-VA3(R)"         "PB(R9)"            "SAD"              
#> [145] "gL(R)"             "CA(L)"             "a2(R)"            
#> [148] "SAD(-AMMC)"        "FBl2"              "AL-D(R)"          
#> [151] "AL-VM5v(R)"        "SMP(R)"            "AL-VM7v(R)"       
#> [154] "FBl7"              "FBl9"              "AL-DM6(R)"        
#> [157] "AL-V(R)"           "PB(L3)"            "PB(R1)"           
#> [160] "CA(R)"             "AL-VA7m(R)"        "AL-VC5(R)"        
#> [163] "EBr3d"             "AL-VA1v(R)"        "AL-VA7l(R)"       
#> [166] "AL-VC4(R)"         "PRW"               "ICL(L)"           
#> [169] "SNP(L)"            "FBl8"              "CX"               
#> [172] "AB(R)"             "AL-DM3(L)"         "PB(L6)"           
#> [175] "RUB(L)"            "AL-DC4(R)"         "LX(R)"            
#> [178] "ME(R)"             "WED(R)"            "g2(R)"            
#> [181] "b1(R)"             "AME(R)"            "AL-VP4(R)"        
#> [184] "AL-DP1l(R)"        "AL-VA4(R)"         "PENP"             
#> [187] "hemibrain"         "AL-VA6(L)"         "NO3(L)"           
#> [190] "bL(R)"             "a'L(L)"            "mALT(L)"          
#> [193] "PB(R2)"            "ATL(L)"            "NO2(L)"           
#> [196] "LAL(-GA)(R)"       "PB(L7)"            "AL-DA2(R)"        
#> [199] "CRE(-ROB,-RUB)(R)" "g3(R)"             "NO3(R)"           
#> [202] "EBr3pw"            "LX(L)"             "SLP(R)"           
#> [205] "PB(R7)"            "SCL(R)"            "a'1(R)"           
#> [208] "AL-DL5(R)"         "AL-DM2(L)"         "SIP(R)"           
#> [211] "AL-DC2(R)"         "AL-VA5(R)"         "FB"               
#> [214] "AL-D(L)"           "AL-DC3(R)"         "AL-VA6(R)"        
#> [217] "AL-VP1m(R)"        "AOT(R)"            "AL-DC4(L)"        
#> [220] "PB(R6)"            "FBl5"              "AL-DC1(R)"        
#> [223] "AL-VP1d(R)"        "GC"                "b'L(R)"           
#> [226] "FBl4"              "NO(L)"             "VLNP(R)"          
#> [229] "PB"                "AL(L)"             "PB(L2)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "FB"                "NO1(R)"            "EBr2r4"           
#>   [4] "gL(L)"             "SNP(L)"            "LH(R)"            
#>   [7] "lACA(R)"           "NO1(L)"            "PB(R5)"           
#>  [10] "CX"                "g5(R)"             "SAD(-AMMC)"       
#>  [13] "INP"               "AB(L)"             "LAL(L)"           
#>  [16] "CRE(R)"            "PRW"               "RUB(L)"           
#>  [19] "IPS(R)"            "MB(R)"             "g1(R)"            
#>  [22] "FBl3"              "CA(R)"             "VES(R)"           
#>  [25] "FBl1"              "BU(L)"             "PB(L8)"           
#>  [28] "PB(R1)"            "PB(L7)"            "POC"              
#>  [31] "b'1(R)"            "PENP"              "a2(R)"            
#>  [34] "vACA(R)"           "ICL(R)"            "PB(R4)"           
#>  [37] "SMP(R)"            "NO2(L)"            "MB(L)"            
#>  [40] "EPA(R)"            "CAN(R)"            "aL(L)"            
#>  [43] "SLP(R)"            "CA(L)"             "EBr1"             
#>  [46] "FBl9"              "EBr3d"             "FBl4"             
#>  [49] "SIP(L)"            "ME(R)"             "NO"               
#>  [52] "FBl7"              "SCL(L)"            "PB(L3)"           
#>  [55] "a'L(R)"            "GOR(L)"            "a'3(R)"           
#>  [58] "g4(R)"             "NO3(L)"            "SAD"              
#>  [61] "SNP(R)"            "PB(L6)"            "PB(L9)"           
#>  [64] "LO(R)"             "AB(R)"             "hemibrain"        
#>  [67] "PB"                "b'L(L)"            "LOP(R)"           
#>  [70] "PVLP(R)"           "a3(R)"             "AOT(R)"           
#>  [73] "mALT(L)"           "PLP(R)"            "ICL(L)"           
#>  [76] "VLNP(R)"           "NO3(R)"            "bL(R)"            
#>  [79] "NO(R)"             "EBr3pw"            "AME(R)"           
#>  [82] "AOTU(R)"           "BU(R)"             "SCL(R)"           
#>  [85] "EBr5"              "PB(R9)"            "EBr3am"           
#>  [88] "SMP(L)"            "PED(R)"            "LX(R)"            
#>  [91] "bL(L)"             "VES(L)"            "AL(R)"            
#>  [94] "IB"                "PB(L2)"            "SPS(L)"           
#>  [97] "VMNP"              "CRE(-ROB,-RUB)(R)" "AMMC"             
#> [100] "PB(L1)"            "AL(L)"             "EBr6"             
#> [103] "PB(L5)"            "SPS(R)"            "b2(R)"            
#> [106] "FBl2"              "PB(R2)"            "PB(R3)"           
#> [109] "AVLP(R)"           "LX(L)"             "NO(L)"            
#> [112] "g2(R)"             "CRE(L)"            "PB(L4)"           
#> [115] "PB(R7)"            "LAL(-GA)(R)"       "a1(R)"            
#> [118] "g3(R)"             "GA(R)"             "NO2(R)"           
#> [121] "a'L(L)"            "b'L(R)"            "ATL(R)"           
#> [124] "GOR(R)"            "mALT(R)"           "EPA(L)"           
#> [127] "SIP(R)"            "b1(R)"             "GNG"              
#> [130] "LAL(R)"            "PB(R6)"            "FLA(R)"           
#> [133] "FBl8"              "OL(R)"             "a'1(R)"           
#> [136] "FBl5"              "dACA(R)"           "PB(R8)"           
#> [139] "RUB(R)"            "GC"                "WED(R)"           
#> [142] "gL(R)"             "FBl6"              "a'2(R)"           
#> [145] "ROB(R)"            "ATL(L)"            "GF(R)"            
#> [148] "EB"                "MB(+ACA)(R)"       "b'2(R)"           
#> [151] "aL(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-VC2(R)"         "AL-VC4(R)"         "AL-VL1(R)"        
#>   [4] "AL-DA4m(L)"        "FBl9"              "POC"              
#>   [7] "PB(L7)"            "INP"               "NO(L)"            
#>  [10] "AL-DL1(R)"         "AL-DP1l(R)"        "g1(R)"            
#>  [13] "gL(L)"             "EBr3pw"            "AL-VP3(R)"        
#>  [16] "g3(R)"             "LX(R)"             "SCL(L)"           
#>  [19] "AL-DL5(L)"         "a'3(R)"            "NO1(L)"           
#>  [22] "PVLP(R)"           "AVLP(R)"           "PB(L9)"           
#>  [25] "ATL(R)"            "mALT(L)"           "AL(L)"            
#>  [28] "AL-VA3(R)"         "AL-VM5d(R)"        "LAL(L)"           
#>  [31] "AL-DM4(R)"         "ATL(L)"            "SIP(L)"           
#>  [34] "NO(R)"             "FBl2"              "PB(L2)"           
#>  [37] "b'2(R)"            "CRE(-RUB)(L)"      "AL-DM6(R)"        
#>  [40] "LH(R)"             "WED(R)"            "EBr5"             
#>  [43] "AL-DL4(L)"         "a3(R)"             "AL-VA1d(R)"       
#>  [46] "PB(R2)"            "AL-DA2(R)"         "GA(R)"            
#>  [49] "b2(R)"             "AL-VM7v(L)"        "AOTU(R)"          
#>  [52] "AMMC"              "EBr3am"            "a1(R)"            
#>  [55] "AL-VP1l(R)"        "AL-DL2v(R)"        "vACA(R)"          
#>  [58] "AL-VA6(R)"         "AL-VC3l(R)"        "MB(L)"            
#>  [61] "AL-DC2(L)"         "VES(R)"            "FBl5"             
#>  [64] "CA(R)"             "AL-DA4m(R)"        "AL-DM3(L)"        
#>  [67] "AL-VA7m(R)"        "AB(L)"             "CRE(-ROB,-RUB)(R)"
#>  [70] "SMP(L)"            "CRE(R)"            "VMNP"             
#>  [73] "AL-VL2a(R)"        "aL(R)"             "PB(R3)"           
#>  [76] "AL-DM1(R)"         "EPA(L)"            "OL(R)"            
#>  [79] "PB(L5)"            "AL-DA3(L)"         "bL(L)"            
#>  [82] "b1(R)"             "AL-DM2(L)"         "aL(L)"            
#>  [85] "EBr1"              "BU(R)"             "AL-DP1m(L)"       
#>  [88] "PB(R6)"            "AL-VM4(R)"         "AME(R)"           
#>  [91] "AL-DP1m(R)"        "AL-VM1(R)"         "BU(L)"            
#>  [94] "EBr3d"             "AB(R)"             "FBl8"             
#>  [97] "RUB(L)"            "AL-DC4(L)"         "PB(L6)"           
#> [100] "g5(R)"             "AL-VC5(R)"         "GOR(L)"           
#> [103] "AL-DM6(L)"         "LO(R)"             "AL-DM4(L)"        
#> [106] "SAD(-AMMC)"        "PB(L4)"            "SLP(R)"           
#> [109] "AL-DC4(R)"         "AL-VM7d(R)"        "NO2(L)"           
#> [112] "SAD"               "FBl4"              "FBl6"             
#> [115] "a'1(R)"            "bL(R)"             "VLNP(R)"          
#> [118] "SCL(R)"            "PB(L8)"            "SNP(R)"           
#> [121] "AL-VM5v(R)"        "AL-VA6(L)"         "AL-VA7l(R)"       
#> [124] "GF(R)"             "a2(R)"             "PRW"              
#> [127] "PENP"              "MB(+ACA)(R)"       "AL-VA4(R)"        
#> [130] "AL-VM3(R)"         "AL-VP5(R)"         "ICL(R)"           
#> [133] "AL-DA4l(R)"        "AL-DM5(R)"         "PED(R)"           
#> [136] "ROB(R)"            "gL(R)"             "a'L(R)"           
#> [139] "AL-DC1(L)"         "NO"                "CAN(R)"           
#> [142] "SPS(L)"            "a'2(R)"            "NO3(L)"           
#> [145] "SNP(L)"            "LX(L)"             "AL-VP4(R)"        
#> [148] "CA(L)"             "PB(R8)"            "EBr2r4"           
#> [151] "EPA(R)"            "AL-VC3m(R)"        "AL-VM7v(R)"       
#> [154] "FBl1"              "MB(R)"             "AL-DL3(R)"        
#> [157] "b'L(R)"            "PLP(R)"            "SIP(R)"           
#> [160] "PB(R1)"            "AL-VP1d(R)"        "AL-VA2(R)"        
#> [163] "SPS(R)"            "PB(R7)"            "AL-DL5(R)"        
#> [166] "EB"                "AL-DC2(R)"         "AL-VA5(R)"        
#> [169] "AOT(R)"            "mALT(R)"           "EBr6"             
#> [172] "NO2(R)"            "PB(R9)"            "AL-DL2d(R)"       
#> [175] "AL-DM3(R)"         "AL-VL2p(R)"        "SMP(R)"           
#> [178] "g4(R)"             "AL-D(R)"           "AL-VM2(R)"        
#> [181] "AL-DM5(L)"         "GC"                "FBl7"             
#> [184] "b'1(R)"            "PB(R5)"            "LAL(-GA)(R)"      
#> [187] "AL-D(L)"           "AL-DM2(R)"         "AL-DC1(R)"        
#> [190] "FB"                "g2(R)"             "PB(L1)"           
#> [193] "AL-V(R)"           "AL-DM1(L)"         "IPS(R)"           
#> [196] "b'L(L)"            "AL-VP2(R)"         "CX"               
#> [199] "lACA(R)"           "PB"                "AL-VP1m(R)"       
#> [202] "dACA(R)"           "IB"                "LOP(R)"           
#> [205] "GOR(R)"            "AL-DL4(R)"         "PB(L3)"           
#> [208] "NO1(R)"            "CRE(L)"            "AL-DC3(R)"        
#> [211] "AL-DA2(L)"         "FB-column3"        "VES(L)"           
#> [214] "AL-VA1v(R)"        "a'L(L)"            "AL(R)"            
#> [217] "AL-DA3(R)"         "AL-VM7d(L)"        "PB(R4)"           
#> [220] "GNG"               "AL-VC1(R)"         "hemibrain"        
#> [223] "NO3(R)"            "ME(R)"             "LAL(R)"           
#> [226] "AL-DA1(R)"         "RUB(R)"            "FLA(R)"           
#> [229] "ICL(L)"            "FBl3"             
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "ICL(L)"            "LAL(R)"            "AL-DC1(R)"        
#>   [4] "AL-VM3(R)"         "AL-VP1d(R)"        "a2(R)"            
#>   [7] "AL-VM5d(R)"        "PB(R5)"            "bL(R)"            
#>  [10] "a'L(L)"            "AL-DP1m(R)"        "SAD(-AMMC)"       
#>  [13] "FBl5"              "LH(R)"             "AL-VM4(R)"        
#>  [16] "LAL(L)"            "AL-DC2(L)"         "PB(L6)"           
#>  [19] "AL(R)"             "AL-VA5(R)"         "LAL(-GA)(R)"      
#>  [22] "gL(R)"             "AL-VM7v(R)"        "SPS(R)"           
#>  [25] "AL-VC2(R)"         "b'L(R)"            "AL-DM6(L)"        
#>  [28] "CRE(L)"            "FBl2"              "FBl3"             
#>  [31] "PB(R3)"            "AL-DL4(R)"         "AL-VC4(R)"        
#>  [34] "PB(L3)"            "PB(L1)"            "AL-DM1(L)"        
#>  [37] "PB(R9)"            "CRE(-RUB)(L)"      "NO3(L)"           
#>  [40] "AL-VP1m(R)"        "a'3(R)"            "GF(R)"            
#>  [43] "NO1(R)"            "SPS(L)"            "AL-DA2(R)"        
#>  [46] "AL-DL3(R)"         "AL-VP1l(R)"        "GA(R)"            
#>  [49] "ME(R)"             "FBl9"              "AL-DM1(R)"        
#>  [52] "EB"                "PB(L7)"            "AL-VP2(R)"        
#>  [55] "PENP"              "NO1(L)"            "AL-DP1m(L)"       
#>  [58] "ATL(R)"            "AL-VL2p(R)"        "FB"               
#>  [61] "AOT(R)"            "AVLP(R)"           "PB(L4)"           
#>  [64] "AL-DC2(R)"         "EBr6"              "bL(L)"            
#>  [67] "FBl1"              "FBl4"              "a3(R)"            
#>  [70] "POC"               "hemibrain"         "PB(R1)"           
#>  [73] "NO2(L)"            "AL-DA1(R)"         "AL-DA4m(R)"       
#>  [76] "EBr3am"            "EPA(R)"            "IB"               
#>  [79] "PB(R6)"            "AL-DC1(L)"         "AL-DL2v(R)"       
#>  [82] "AL-VM5v(R)"        "PB"                "AL(L)"            
#>  [85] "FB-column3"        "NO2(R)"            "ATL(L)"           
#>  [88] "SCL(R)"            "VMNP"              "RUB(L)"           
#>  [91] "b'L(L)"            "AL-DM3(R)"         "PB(L2)"           
#>  [94] "dACA(R)"           "LX(L)"             "AL-VM7d(R)"       
#>  [97] "mALT(L)"           "NO"                "ICL(R)"           
#> [100] "INP"               "AL-DP1l(R)"        "AL-VA7m(R)"       
#> [103] "AL-VL1(R)"         "mALT(R)"           "AL-VA7l(R)"       
#> [106] "VLNP(R)"           "SMP(L)"            "MB(L)"            
#> [109] "SNP(R)"            "LOP(R)"            "PRW"              
#> [112] "SIP(R)"            "CA(R)"             "AL-VM1(R)"        
#> [115] "AMMC"              "LX(R)"             "AL-DA2(L)"        
#> [118] "SMP(R)"            "SCL(L)"            "SAD"              
#> [121] "a'2(R)"            "AL-DL5(L)"         "MB(+ACA)(R)"      
#> [124] "AL-VP4(R)"         "RUB(R)"            "CRE(R)"           
#> [127] "g2(R)"             "g4(R)"             "AL-DM6(R)"        
#> [130] "AL-VA4(R)"         "AL-VA6(R)"         "AL-DM2(R)"        
#> [133] "a'L(R)"            "AL-DM2(L)"         "EBr3pw"           
#> [136] "g5(R)"             "AL-DM5(R)"         "AL-VA6(L)"        
#> [139] "AME(R)"            "b'2(R)"            "g3(R)"            
#> [142] "lACA(R)"           "WED(R)"            "b2(R)"            
#> [145] "a'1(R)"            "gL(L)"             "AL-DA4m(L)"       
#> [148] "NO(L)"             "LO(R)"             "CAN(R)"           
#> [151] "GOR(L)"            "EBr2r4"            "MB(R)"            
#> [154] "VES(R)"            "b'1(R)"            "CRE(-ROB,-RUB)(R)"
#> [157] "BU(L)"             "AL-VM7v(L)"        "IPS(R)"           
#> [160] "AL-DL2d(R)"        "PED(R)"            "NO(R)"            
#> [163] "GNG"               "AL-DC3(R)"         "AL-VC3m(R)"       
#> [166] "CA(L)"             "AOTU(R)"           "AL-DL1(R)"        
#> [169] "AL-VL2a(R)"        "a1(R)"             "AL-VC1(R)"        
#> [172] "AL-DL5(R)"         "AL-VA1v(R)"        "AL-VA3(R)"        
#> [175] "PB(L5)"            "AL-DA3(L)"         "aL(L)"            
#> [178] "AL-D(R)"           "EBr5"              "PB(L8)"           
#> [181] "AL-VC5(R)"         "AL-VM2(R)"         "AL-VP5(R)"        
#> [184] "GC"                "g1(R)"             "PB(R2)"           
#> [187] "AL-V(R)"           "AB(L)"             "BU(R)"            
#> [190] "SNP(L)"            "VES(L)"            "AL-DM4(R)"        
#> [193] "EBr3d"             "AL-DA4l(R)"        "FLA(R)"           
#> [196] "PB(R8)"            "AL-DL4(L)"         "AL-DC4(R)"        
#> [199] "EBr1"              "SLP(R)"            "AL-VP3(R)"        
#> [202] "AL-DM3(L)"         "EPA(L)"            "OL(R)"            
#> [205] "AL-VA1d(R)"        "PB(R7)"            "AL-D(L)"          
#> [208] "AL-DM4(L)"         "AL-DM5(L)"         "PVLP(R)"          
#> [211] "FBl8"              "FBl7"              "AB(R)"            
#> [214] "AL-DA3(R)"         "NO3(R)"            "ROB(R)"           
#> [217] "PB(R4)"            "SIP(L)"            "aL(R)"            
#> [220] "PLP(R)"            "AL-DC4(L)"         "PB(L9)"           
#> [223] "FBl6"              "GOR(R)"            "AL-VA2(R)"        
#> [226] "vACA(R)"           "CX"                "AL-VM7d(L)"       
#> [229] "b1(R)"             "AL-VC3l(R)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                           
#> [1,] "POST" "/api/v:ver/raw/keyvalue/key/:instance/:key"   
#> [2,] "GET"  "/api/v:ver/npexplorer/autapses"               
#> [3,] "POST" "/api/v:ver/npexplorer/celltype/:dataset/:type"
#> [4,] "POST" "/api/raw/cypher/transaction/:id/kill"         
#> [5,] "GET"  "/api/npexplorer/rankedtable"                  
#> [6,] "POST" "/api/npexplorer/commonconnectivity"           
#>      name                                                                                              
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                          
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getRankedTable-fm"       
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCommonConnectivity-fm"
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }