Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "NO(L)"             "PENP"              "g3(R)"            
#>   [4] "SCL(R)"            "NO1(R)"            "EPA(L)"           
#>   [7] "VMNP"              "LAL(R)"            "SCL(L)"           
#>  [10] "GC"                "MB(L)"             "SNP(L)"           
#>  [13] "LH(R)"             "AMMC"              "PB(R9)"           
#>  [16] "MB(+ACA)(R)"       "AB(R)"             "AB(L)"            
#>  [19] "vACA(R)"           "EBr3d"             "a'L(R)"           
#>  [22] "FBl3"              "SIP(L)"            "AME(R)"           
#>  [25] "PB(R5)"            "CRE(-ROB,-RUB)(R)" "PB(L8)"           
#>  [28] "g4(R)"             "NO(R)"             "AVLP(R)"          
#>  [31] "VLNP(R)"           "ME(R)"             "FBl9"             
#>  [34] "ATL(R)"            "b1(R)"             "AL(R)"            
#>  [37] "a1(R)"             "LAL(L)"            "PB(R7)"           
#>  [40] "ICL(R)"            "PED(R)"            "LO(R)"            
#>  [43] "LOP(R)"            "lACA(R)"           "PB(L6)"           
#>  [46] "PB(R1)"            "bL(L)"             "FBl8"             
#>  [49] "PB(R4)"            "OL(R)"             "hemibrain"        
#>  [52] "EBr6"              "PB"                "SNP(R)"           
#>  [55] "gL(L)"             "FBl5"              "EBr3am"           
#>  [58] "PRW"               "PB(L1)"            "aL(L)"            
#>  [61] "a'3(R)"            "CRE(R)"            "PB(L9)"           
#>  [64] "EBr3pw"            "PB(R3)"            "CX"               
#>  [67] "NO3(R)"            "FLA(R)"            "LX(R)"            
#>  [70] "BU(R)"             "FBl6"              "NO2(R)"           
#>  [73] "b'L(L)"            "EBr2r4"            "ROB(R)"           
#>  [76] "a'1(R)"            "GOR(R)"            "PB(L4)"           
#>  [79] "PB(R6)"            "SIP(R)"            "gL(R)"            
#>  [82] "VES(L)"            "b'2(R)"            "SPS(R)"           
#>  [85] "CRE(L)"            "g5(R)"             "PB(R2)"           
#>  [88] "RUB(R)"            "FB"                "a'2(R)"           
#>  [91] "PVLP(R)"           "a2(R)"             "EBr1"             
#>  [94] "FBl7"              "CAN(R)"            "RUB(L)"           
#>  [97] "SMP(L)"            "LAL(-GA)(R)"       "b2(R)"            
#> [100] "EBr5"              "GNG"               "AOT(R)"           
#> [103] "GF(R)"             "WED(R)"            "SPS(L)"           
#> [106] "a'L(L)"            "POC"               "mALT(R)"          
#> [109] "BU(L)"             "SMP(R)"            "CA(L)"            
#> [112] "g1(R)"             "IPS(R)"            "PB(L5)"           
#> [115] "MB(R)"             "AOTU(R)"           "dACA(R)"          
#> [118] "NO"                "ICL(L)"            "PB(L7)"           
#> [121] "LX(L)"             "SLP(R)"            "NO2(L)"           
#> [124] "NO1(L)"            "FBl2"              "FBl4"             
#> [127] "a3(R)"             "mALT(L)"           "PB(L3)"           
#> [130] "EB"                "EPA(R)"            "SAD(-AMMC)"       
#> [133] "CA(R)"             "PB(L2)"            "ATL(L)"           
#> [136] "GOR(L)"            "g2(R)"             "AL(L)"            
#> [139] "bL(R)"             "IB"                "SAD"              
#> [142] "FBl1"              "GA(R)"             "b'L(R)"           
#> [145] "INP"               "b'1(R)"            "aL(R)"            
#> [148] "PB(R8)"            "PLP(R)"            "NO3(L)"           
#> [151] "VES(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                      
#> [1,] "GET"  "/api/custom/custom"                      
#> [2,] "POST" "/api/npexplorer/rankedtable"             
#> [3,] "GET"  "/api/v:ver/npexplorer/commonconnectivity"
#> [4,] "GET"  "/api/v:ver/custom/custom"                
#> [5,] "GET"  "/api/v:ver/npexplorer/neuronmetavals"    
#> [6,] "GET"  "/api/npexplorer/roiconnectivity"         
#>      name                                                                                           
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"                
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getRankedTable-fm"    
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIConnectivity-fm"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }