Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "FB-column3"        "ATL(R)"            "AL-DM1(R)"        
#>   [4] "AL-VM5v(R)"        "AL-VP2(R)"         "PED(R)"           
#>   [7] "AL-DL4(L)"         "AL-DL3(R)"         "AL-VA4(R)"        
#>  [10] "EBr3am"            "AL-DM2(L)"         "PB(L6)"           
#>  [13] "NO(R)"             "a'1(R)"            "ME(R)"            
#>  [16] "g4(R)"             "AL-VM4(R)"         "AL-D(L)"          
#>  [19] "SLP(R)"            "AL-DC2(R)"         "AL-DC3(R)"        
#>  [22] "AL-D(R)"           "AL-VM3(R)"         "CRE(R)"           
#>  [25] "BU(L)"             "EPA(L)"            "AOTU(R)"          
#>  [28] "RUB(L)"            "GOR(R)"            "AL-DM5(R)"        
#>  [31] "AMMC"              "AL-DM1(L)"         "PB(L7)"           
#>  [34] "IB112"             "AL-VA1d(R)"        "a'L(L)"           
#>  [37] "OL(R)"             "AL-VP5(R)"         "a1(R)"            
#>  [40] "g3(R)"             "GA(R)"             "b'L(L)"           
#>  [43] "AL-DM6(R)"         "AL-VC5(R)"         "AL-VM2(R)"        
#>  [46] "CX"                "g2(R)"             "EBr5"             
#>  [49] "AL-DC4(R)"         "b2(R)"             "AL-DC4(L)"        
#>  [52] "PB(R1)"            "ICL(R)"            "AL-DA2(R)"        
#>  [55] "AL-VA7l(R)"        "AL-VM7v(R)"        "hemibrain"        
#>  [58] "AL-DA2(L)"         "AL-DA4m(L)"        "GNG"              
#>  [61] "RUB(R)"            "aL(R)"             "INP"              
#>  [64] "EBr2r4"            "ICL(L)"            "AL-DP1l(R)"       
#>  [67] "AL-VP4(R)"         "mALT(L)"           "PRW"              
#>  [70] "PB(R9)"            "PB(L9)"            "AL-DL1(R)"        
#>  [73] "AL-VA2(R)"         "SMP(L)"            "ROB(R)"           
#>  [76] "g5(R)"             "CA(R)"             "AME(R)"           
#>  [79] "MB(L)"             "PB(R5)"            "PENP"             
#>  [82] "FBl7"              "LH(R)"             "AL-DL5(R)"        
#>  [85] "a'3(R)"            "MB(R)"             "AL-VA3(R)"        
#>  [88] "PB(R2)"            "AL-DL5(L)"         "AL-VP1l(R)"       
#>  [91] "SMP(R)"            "b'1(R)"            "PLP(R)"           
#>  [94] "FBl6"              "AVLP(R)"           "NO(L)"            
#>  [97] "AL-VC3m(R)"        "vACA(R)"           "FB"               
#> [100] "bL(R)"             "AL-DL4(R)"         "AL-VC2(R)"        
#> [103] "LAL(R)"            "a'2(R)"            "NO2(R)"           
#> [106] "AL-VM7d(R)"        "AB(L)"             "NO3(R)"           
#> [109] "SCL(R)"            "VMNP"              "gL(R)"            
#> [112] "LX(L)"             "PB(L1)"            "SPS(L)"           
#> [115] "AL-DP1m(L)"        "PB(L3)"            "g1(R)"            
#> [118] "SIP(R)"            "AB(R)"             "AL-DM2(R)"        
#> [121] "AL-VL2a(R)"        "EBr3pw"            "FBl2"             
#> [124] "AL-DA1(R)"         "SNP(R)"            "AL-DM4(L)"        
#> [127] "AL-DM5(L)"         "CAN(R)"            "AL(R)"            
#> [130] "FBl1"              "SCL(L)"            "LAL(L)"           
#> [133] "b'L(R)"            "NO1(R)"            "IB"               
#> [136] "NO2(L)"            "EBr1"              "PB(R4)"           
#> [139] "AL-DP1m(R)"        "AL-VA7m(R)"        "AOT(R)"           
#> [142] "AL-DC2(L)"         "CA(L)"             "CRE(-RUB)(L)"     
#> [145] "AL-VM5d(R)"        "SAD"               "PB(R3)"           
#> [148] "AL-VC4(R)"         "AL-VM1(R)"         "AL-V(R)"          
#> [151] "AL-VP1m(R)"        "GC"                "GOR(L)"           
#> [154] "FBl3"              "AL-DL2d(R)"        "AL-DC1(L)"        
#> [157] "AL-DC1(R)"         "CRE(L)"            "SNP(L)"           
#> [160] "FBl5"              "AL-DA4l(R)"        "a'L(R)"           
#> [163] "AL-DL2v(R)"        "PB(L8)"            "EBr6"             
#> [166] "bL(L)"             "VES(R)"            "aL(L)"            
#> [169] "b'2(R)"            "PB(L4)"            "EB"               
#> [172] "AL-VC1(R)"         "AL-VP1d(R)"        "EBr3d"            
#> [175] "gL(L)"             "VES(L)"            "GF(R)"            
#> [178] "dACA(R)"           "NO3(L)"            "AL-VL1(R)"        
#> [181] "lACA(R)"           "AL-VL2p(R)"        "NO1(L)"           
#> [184] "AL-DM3(L)"         "AL-VA6(R)"         "VLNP(R)"          
#> [187] "POC"               "SIP(L)"            "PB(R7)"           
#> [190] "PB"                "BU(R)"             "AL-VA6(L)"        
#> [193] "AL-VM7d(L)"        "AL-DM4(R)"         "LOP(R)"           
#> [196] "a3(R)"             "AL-VC3l(R)"        "LAL(-GA)(R)"      
#> [199] "WED(R)"            "EPA(R)"            "NO"               
#> [202] "mALT(R)"           "LO(R)"             "MB(+ACA)(R)"      
#> [205] "b1(R)"             "IPS(R)"            "a2(R)"            
#> [208] "LX(R)"             "AL-VA5(R)"         "PB(R6)"           
#> [211] "FLA(R)"            "AL-DM6(L)"         "PVLP(R)"          
#> [214] "FBl4"              "FBl8"              "AL-DA3(R)"        
#> [217] "SPS(R)"            "AL(L)"             "AL-DA3(L)"        
#> [220] "SAD(-AMMC)"        "AL-DM3(R)"         "PB(L5)"           
#> [223] "AL-VP3(R)"         "PB(L2)"            "ATL(L)"           
#> [226] "PB(R8)"            "AL-VM7v(L)"        "FBl9"             
#> [229] "AL-DA4m(R)"        "AL-VA1v(R)"        "CRE(-ROB,-RUB)(R)"
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(R4)"            "AOT(R)"            "hemibrain"        
#>   [4] "PED(R)"            "a'1(R)"            "VES(R)"           
#>   [7] "FBl1"              "VLNP(R)"           "SCL(L)"           
#>  [10] "CRE(-ROB,-RUB)(R)" "EBr3d"             "PVLP(R)"          
#>  [13] "b1(R)"             "VES(L)"            "AL(R)"            
#>  [16] "ATL(L)"            "bL(L)"             "FBl4"             
#>  [19] "INP"               "mALT(L)"           "NO2(L)"           
#>  [22] "SNP(L)"            "ATL(R)"            "PENP"             
#>  [25] "PB(L3)"            "SAD"               "PB(R7)"           
#>  [28] "WED(R)"            "aL(L)"             "PB(L2)"           
#>  [31] "NO3(R)"            "NO"                "NO(L)"            
#>  [34] "NO(R)"             "EBr3pw"            "MB(+ACA)(R)"      
#>  [37] "GOR(R)"            "lACA(R)"           "EBr1"             
#>  [40] "EBr2r4"            "FBl5"              "AVLP(R)"          
#>  [43] "CA(R)"             "AB(R)"             "bL(R)"            
#>  [46] "PB(L4)"            "LAL(-GA)(R)"       "mALT(R)"          
#>  [49] "AL(L)"             "vACA(R)"           "PB(L9)"           
#>  [52] "a'L(L)"            "MB(R)"             "aL(R)"            
#>  [55] "NO1(R)"            "LX(R)"             "SNP(R)"           
#>  [58] "g4(R)"             "EPA(L)"            "MB(L)"            
#>  [61] "LO(R)"             "FBl2"              "FBl6"             
#>  [64] "AME(R)"            "PB(L6)"            "NO2(R)"           
#>  [67] "NO1(L)"            "g2(R)"             "FLA(R)"           
#>  [70] "PLP(R)"            "gL(L)"             "PB"               
#>  [73] "a3(R)"             "SMP(L)"            "CRE(R)"           
#>  [76] "IPS(R)"            "IB"                "AMMC"             
#>  [79] "GOR(L)"            "g1(R)"             "SAD(-AMMC)"       
#>  [82] "PB(R3)"            "ICL(R)"            "SMP(R)"           
#>  [85] "PB(L1)"            "EB"                "PB(L7)"           
#>  [88] "GNG"               "SLP(R)"            "NO3(L)"           
#>  [91] "g5(R)"             "PB(R2)"            "GF(R)"            
#>  [94] "SCL(R)"            "VMNP"              "a'L(R)"           
#>  [97] "EPA(R)"            "PB(R5)"            "RUB(R)"           
#> [100] "a'2(R)"            "SIP(R)"            "LOP(R)"           
#> [103] "b'L(R)"            "ICL(L)"            "FBl8"             
#> [106] "CAN(R)"            "AB(L)"             "PB(R9)"           
#> [109] "a'3(R)"            "CX"                "PB(L8)"           
#> [112] "FBl7"              "b'2(R)"            "LAL(L)"           
#> [115] "g3(R)"             "a2(R)"             "GC"               
#> [118] "CRE(L)"            "a1(R)"             "AOTU(R)"          
#> [121] "ME(R)"             "dACA(R)"           "gL(R)"            
#> [124] "PB(L5)"            "GA(R)"             "SIP(L)"           
#> [127] "POC"               "b'1(R)"            "b2(R)"            
#> [130] "EBr6"              "BU(L)"             "LAL(R)"           
#> [133] "CA(L)"             "PB(R6)"            "EBr5"             
#> [136] "SPS(L)"            "b'L(L)"            "SPS(R)"           
#> [139] "PB(R1)"            "FBl3"              "FBl9"             
#> [142] "LX(L)"             "OL(R)"             "PRW"              
#> [145] "FB"                "BU(R)"             "ROB(R)"           
#> [148] "LH(R)"             "PB(R8)"            "RUB(L)"           
#> [151] "EBr3am"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "POC"               "CRE(-ROB,-RUB)(R)" "FLA(R)"           
#>   [4] "AME(R)"            "AL-DP1l(R)"        "AB(L)"            
#>   [7] "GNG"               "AL-DL5(R)"         "AL-VA2(R)"        
#>  [10] "AL-VM3(R)"         "AL-VP2(R)"         "VMNP"             
#>  [13] "AL-DC2(R)"         "AL-DM3(L)"         "AL-VM7d(L)"       
#>  [16] "NO"                "a'1(R)"            "bL(R)"            
#>  [19] "AL-DA2(R)"         "WED(R)"            "BU(R)"            
#>  [22] "a'L(L)"            "AL-DC1(R)"         "LAL(L)"           
#>  [25] "PVLP(R)"           "IB"                "PB(L9)"           
#>  [28] "PB(R3)"            "VES(R)"            "AL-DA3(R)"        
#>  [31] "PLP(R)"            "ICL(L)"            "NO3(R)"           
#>  [34] "NO1(L)"            "AL-D(R)"           "AL-DM4(R)"        
#>  [37] "CRE(-RUB)(L)"      "FBl7"              "SLP(R)"           
#>  [40] "VLNP(R)"           "PENP"              "g4(R)"            
#>  [43] "g5(R)"             "CAN(R)"            "AL-VA6(R)"        
#>  [46] "hemibrain"         "PB(R1)"            "AL-DC2(L)"        
#>  [49] "NO(L)"             "AL-DA4m(L)"        "ICL(R)"           
#>  [52] "PB(R2)"            "AL-DL3(R)"         "PRW"              
#>  [55] "PED(R)"            "EB"                "a'L(R)"           
#>  [58] "AL(L)"             "FBl4"              "AL-VA7l(R)"       
#>  [61] "PB(L1)"            "mALT(L)"           "FBl2"             
#>  [64] "AL-V(R)"           "AOT(R)"            "AMMC"             
#>  [67] "AOTU(R)"           "EBr2r4"            "NO(R)"            
#>  [70] "FBl1"              "a3(R)"             "PB(L2)"           
#>  [73] "SPS(R)"            "b'L(R)"            "VES(L)"           
#>  [76] "NO2(L)"            "IPS(R)"            "INP"              
#>  [79] "AL-VA7m(R)"        "a1(R)"             "mALT(R)"          
#>  [82] "EBr5"              "AL-VA5(R)"         "CX"               
#>  [85] "AL-DC1(L)"         "gL(R)"             "bL(L)"            
#>  [88] "AL(R)"             "AL-VC4(R)"         "AL-VM4(R)"        
#>  [91] "g1(R)"             "AL-DA3(L)"         "AL-VM7v(R)"       
#>  [94] "AL-VP3(R)"         "PB"                "AL-VA1d(R)"       
#>  [97] "AL-VM5v(R)"        "AL-VC3m(R)"        "AL-VL2p(R)"       
#> [100] "PB(L4)"            "PB(R5)"            "PB(L5)"           
#> [103] "ATL(L)"            "SIP(R)"            "aL(L)"            
#> [106] "AL-DL4(L)"         "AL-VM7v(L)"        "LH(R)"            
#> [109] "GOR(R)"            "AL-DL4(R)"         "b'2(R)"           
#> [112] "dACA(R)"           "SAD(-AMMC)"        "LX(L)"            
#> [115] "MB(L)"             "AL-DC4(L)"         "AL-DM1(R)"        
#> [118] "AL-DM3(R)"         "FBl9"              "EPA(L)"           
#> [121] "SAD"               "AL-DM4(L)"         "AL-VP4(R)"        
#> [124] "SMP(L)"            "gL(L)"             "AL-DA1(R)"        
#> [127] "AL-DL2v(R)"        "b'1(R)"            "EBr1"             
#> [130] "ROB(R)"            "SNP(R)"            "AL-DL2d(R)"       
#> [133] "PB(L3)"            "EBr3d"             "AL-DM5(L)"        
#> [136] "AL-DC4(R)"         "SMP(R)"            "lACA(R)"          
#> [139] "EBr6"              "AL-VP1m(R)"        "PB(R8)"           
#> [142] "FBl8"              "AL-VP1d(R)"        "AL-DP1m(R)"       
#> [145] "AL-VP5(R)"         "PB(L7)"            "LAL(R)"           
#> [148] "SCL(L)"            "SIP(L)"            "AL-DL5(L)"        
#> [151] "SCL(R)"            "AL-D(L)"           "AL-VA3(R)"        
#> [154] "AL-VC3l(R)"        "NO1(R)"            "ATL(R)"           
#> [157] "PB(L6)"            "OL(R)"             "SPS(L)"           
#> [160] "LX(R)"             "PB(R4)"            "AL-DA4l(R)"       
#> [163] "AL-VC5(R)"         "b2(R)"             "AVLP(R)"          
#> [166] "AL-DC3(R)"         "LOP(R)"            "a'3(R)"           
#> [169] "vACA(R)"           "ME(R)"             "RUB(L)"           
#> [172] "AL-VM1(R)"         "g3(R)"             "NO3(L)"           
#> [175] "AL-VM7d(R)"        "GC"                "FBl5"             
#> [178] "BU(L)"             "RUB(R)"            "AL-DM6(R)"        
#> [181] "PB(R7)"            "LO(R)"             "EBr3am"           
#> [184] "GF(R)"             "a'2(R)"            "AL-VA4(R)"        
#> [187] "CRE(L)"            "AL-VA1v(R)"        "AL-VL2a(R)"       
#> [190] "AL-DP1m(L)"        "AL-VC2(R)"         "EBr3pw"           
#> [193] "AL-DL1(R)"         "AL-VC1(R)"         "PB(R6)"           
#> [196] "aL(R)"             "NO2(R)"            "AL-DA4m(R)"       
#> [199] "AL-VL1(R)"         "EPA(R)"            "AB(R)"            
#> [202] "SNP(L)"            "FB-column3"        "MB(R)"            
#> [205] "a2(R)"             "GOR(L)"            "AL-DM5(R)"        
#> [208] "CRE(R)"            "AL-DM6(L)"         "g2(R)"            
#> [211] "PB(L8)"            "PB(R9)"            "MB(+ACA)(R)"      
#> [214] "b1(R)"             "LAL(-GA)(R)"       "FB"               
#> [217] "AL-VP1l(R)"        "AL-DM2(L)"         "AL-VA6(L)"        
#> [220] "FBl6"              "AL-VM2(R)"         "b'L(L)"           
#> [223] "GA(R)"             "AL-DM1(L)"         "AL-DM2(R)"        
#> [226] "AL-VM5d(R)"        "CA(R)"             "CA(L)"            
#> [229] "AL-DA2(L)"         "FBl3"             
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-DM4(R)"         "AL-VA7m(R)"        "AL-VM7d(R)"       
#>   [4] "AL-DC1(L)"         "INP"               "BU(L)"            
#>   [7] "AL-DA4l(R)"        "NO3(L)"            "AL-DM3(R)"        
#>  [10] "a'2(R)"            "FBl2"              "AL-DL3(R)"        
#>  [13] "hemibrain"         "FLA(R)"            "GA(R)"            
#>  [16] "AL-DC3(R)"         "POC"               "PB(L2)"           
#>  [19] "CAN(R)"            "AL-DC2(R)"         "AL-VA7l(R)"       
#>  [22] "PB(L1)"            "NO1(R)"            "AL-VP1m(R)"       
#>  [25] "AL-DL5(L)"         "AL-DM6(L)"         "VES(L)"           
#>  [28] "NO2(L)"            "WED(R)"            "b2(R)"            
#>  [31] "AL-DM3(L)"         "AL-V(R)"           "SIP(L)"           
#>  [34] "GC"                "AL-DP1m(L)"        "SCL(L)"           
#>  [37] "g3(R)"             "AL-DM5(L)"         "LH(R)"            
#>  [40] "AB(R)"             "RUB(L)"            "PB(L9)"           
#>  [43] "g5(R)"             "AL-VL2a(R)"        "SNP(L)"           
#>  [46] "AL-DM2(R)"         "AL-VM1(R)"         "lACA(R)"          
#>  [49] "AL-DC4(L)"         "NO1(L)"            "AL-VP1l(R)"       
#>  [52] "RUB(R)"            "CX"                "PB"               
#>  [55] "LAL(R)"            "AL-DA4m(L)"        "GNG"              
#>  [58] "mALT(L)"           "aL(R)"             "PB(L8)"           
#>  [61] "AL(R)"             "EBr2r4"            "CRE(L)"           
#>  [64] "EB"                "AB(L)"             "ATL(L)"           
#>  [67] "FB"                "EBr3pw"            "AL-DM5(R)"        
#>  [70] "AL-VA6(R)"         "AL-VC4(R)"         "AL-VM4(R)"        
#>  [73] "a'3(R)"            "g1(R)"             "SIP(R)"           
#>  [76] "FB-column3"        "PB(R5)"            "CRE(-RUB)(L)"     
#>  [79] "SLP(R)"            "AL-VC5(R)"         "AL-DM2(L)"        
#>  [82] "GOR(R)"            "AL-VC1(R)"         "AL-DC4(R)"        
#>  [85] "AL-DL2v(R)"        "AL-VL2p(R)"        "PB(R6)"           
#>  [88] "AL-DC1(R)"         "EBr5"              "PB(L6)"           
#>  [91] "LO(R)"             "a'L(L)"            "AL-DA1(R)"        
#>  [94] "GF(R)"             "GOR(L)"            "FBl3"             
#>  [97] "AVLP(R)"           "EPA(L)"            "AOT(R)"           
#> [100] "VMNP"              "FBl1"              "LX(L)"            
#> [103] "MB(R)"             "dACA(R)"           "b1(R)"            
#> [106] "AL-DA4m(R)"        "AL-VA1v(R)"        "CRE(-ROB,-RUB)(R)"
#> [109] "EBr6"              "AL-VM7v(L)"        "PB(R8)"           
#> [112] "a1(R)"             "AL-DM1(L)"         "AL-D(R)"          
#> [115] "AL-VP1d(R)"        "mALT(R)"           "AL(L)"            
#> [118] "SNP(R)"            "EBr3am"            "a2(R)"            
#> [121] "MB(L)"             "a'1(R)"            "aL(L)"            
#> [124] "AL-VA5(R)"         "PB(L3)"            "LAL(-GA)(R)"      
#> [127] "AL-DA2(L)"         "AL-VM7d(L)"        "CA(R)"            
#> [130] "AL-VP5(R)"         "b'2(R)"            "BU(R)"            
#> [133] "AL-D(L)"           "ROB(R)"            "EPA(R)"           
#> [136] "PB(L5)"            "NO(L)"             "PB(L4)"           
#> [139] "AL-VC3l(R)"        "LAL(L)"            "SPS(L)"           
#> [142] "gL(L)"             "AL-VM5d(R)"        "VES(R)"           
#> [145] "AL-VA3(R)"         "NO2(R)"            "PED(R)"           
#> [148] "NO"                "PLP(R)"            "SMP(L)"           
#> [151] "NO3(R)"            "PENP"              "g2(R)"            
#> [154] "PB(R1)"            "AL-DM6(R)"         "AL-VC2(R)"        
#> [157] "VLNP(R)"           "PRW"               "b'L(R)"           
#> [160] "AOTU(R)"           "SPS(R)"            "FBl4"             
#> [163] "bL(R)"             "IB"                "PB(R7)"           
#> [166] "a'L(R)"            "AL-DL5(R)"         "ATL(R)"           
#> [169] "ICL(R)"            "AL-DA3(L)"         "CA(L)"            
#> [172] "AL-VM5v(R)"        "AL-VP4(R)"         "SMP(R)"           
#> [175] "AL-DM4(L)"         "FBl5"              "AL-VL1(R)"        
#> [178] "LOP(R)"            "LX(R)"             "NO(R)"            
#> [181] "PB(R2)"            "IPS(R)"            "AL-DM1(R)"        
#> [184] "AMMC"              "SCL(R)"            "AL-VA6(L)"        
#> [187] "PB(R3)"            "AME(R)"            "ICL(L)"           
#> [190] "a3(R)"             "AL-VM3(R)"         "AL-VP3(R)"        
#> [193] "g4(R)"             "bL(L)"             "b'1(R)"           
#> [196] "PVLP(R)"           "AL-VA2(R)"         "PB(R9)"           
#> [199] "MB(+ACA)(R)"       "OL(R)"             "ME(R)"            
#> [202] "EBr3d"             "AL-VA4(R)"         "AL-DP1l(R)"       
#> [205] "AL-DP1m(R)"        "PB(R4)"            "AL-DL1(R)"        
#> [208] "AL-DL4(R)"         "gL(R)"             "SAD(-AMMC)"       
#> [211] "AL-VM7v(R)"        "FBl6"              "AL-VC3m(R)"       
#> [214] "AL-VA1d(R)"        "SAD"               "AL-DA3(R)"        
#> [217] "b'L(L)"            "AL-DA2(R)"         "AL-DL2d(R)"       
#> [220] "FBl7"              "FBl9"              "AL-VM2(R)"        
#> [223] "AL-VP2(R)"         "vACA(R)"           "AL-DC2(L)"        
#> [226] "PB(L7)"            "AL-DL4(L)"         "FBl8"             
#> [229] "CRE(R)"            "EBr1"             
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                                     
#> [1,] "OPTIONS" "/api"                                   
#> [2,] "REPORT"  "/api/*"                                 
#> [3,] "POST"    "/api/npexplorer/neuronmetavals"         
#> [4,] "POST"    "/api/v:ver/npexplorer/simpleconnections"
#> [5,] "GET"     "/api/cached/roicompleteness"            
#> [6,] "GET"     "/login"                                 
#>      name                                                                                          
#> [1,] "github.com/labstack/echo.glob..func1"                                                        
#> [2,] "github.com/labstack/echo.glob..func1"                                                        
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMetaVals-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROICompleteness-fm"      
#> [6,] "github.com/janelia-flyem/echo-secure.loginHandler"                                           

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }