Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DM6(L)"         "GA(R)"             "PB(L4)"           
#>   [4] "AL-D(R)"           "AL-VC5(R)"         "g1(R)"            
#>   [7] "AL(R)"             "AL-VM4(R)"         "AOTU(R)"          
#>  [10] "AL-DC2(L)"         "AL-VM7d(L)"        "bL(R)"            
#>  [13] "AL-DL2d(R)"        "AL-DP1l(R)"        "a1(R)"            
#>  [16] "PVLP(R)"           "LAL(R)"            "AMMC"             
#>  [19] "LAL(L)"            "FLA(R)"            "AL-DM5(R)"        
#>  [22] "AL-DM6(R)"         "g4(R)"             "AME(R)"           
#>  [25] "GOR(R)"            "AOT(R)"            "LO(R)"            
#>  [28] "AL-VA7l(R)"        "NO"                "gL(R)"            
#>  [31] "CRE(-ROB,-RUB)(R)" "PB(L3)"            "PB(R1)"           
#>  [34] "RUB(L)"            "OL(R)"             "AL-VA6(R)"        
#>  [37] "LH(R)"             "NO2(R)"            "CX"               
#>  [40] "EBr1"              "AL-DM1(R)"         "AL-DM3(L)"        
#>  [43] "FBl4"              "AL-VA5(R)"         "AL-VL2p(R)"       
#>  [46] "AL-VA4(R)"         "AL-VP1d(R)"        "PED(R)"           
#>  [49] "AL-DP1m(L)"        "ICL(R)"            "SCL(L)"           
#>  [52] "AL-D(L)"           "FB-column3"        "AL-VL1(R)"        
#>  [55] "a'L(R)"            "SAD(-AMMC)"        "a'3(R)"           
#>  [58] "CRE(-RUB)(L)"      "AL-DL2v(R)"        "AL-VA1v(R)"       
#>  [61] "AL-DC4(L)"         "PB(R8)"            "a'2(R)"           
#>  [64] "MB(R)"             "AL-VP3(R)"         "IB"               
#>  [67] "AL-VA6(L)"         "AL-DA4l(R)"        "AL-DM2(R)"        
#>  [70] "AL-DA2(L)"         "INP"               "AL-VC2(R)"        
#>  [73] "PB(L5)"            "GOR(L)"            "EBr6"             
#>  [76] "AL-VM7v(L)"        "AL-DM3(R)"         "AL-DC1(L)"        
#>  [79] "SIP(R)"            "AL-DP1m(R)"        "AL-VP4(R)"        
#>  [82] "NO1(L)"            "EPA(R)"            "AL-DL4(R)"        
#>  [85] "PB(L7)"            "PB(R4)"            "SLP(R)"           
#>  [88] "MB(+ACA)(R)"       "AL-VM2(R)"         "AB(L)"            
#>  [91] "b'2(R)"            "SAD"               "EBr2r4"           
#>  [94] "SNP(L)"            "CAN(R)"            "WED(R)"           
#>  [97] "EBr5"              "EBr3d"             "a3(R)"            
#> [100] "NO(L)"             "RUB(R)"            "ATL(R)"           
#> [103] "lACA(R)"           "AL-DC1(R)"         "AL-DC3(R)"        
#> [106] "AL-DM5(L)"         "a'1(R)"            "FBl9"             
#> [109] "PB(L2)"            "g3(R)"             "FB"               
#> [112] "b'L(R)"            "NO(R)"             "AL-DA2(R)"        
#> [115] "AL-VM5d(R)"        "AL-DM4(L)"         "AL-VP5(R)"        
#> [118] "PRW"               "AL-DM1(L)"         "GNG"              
#> [121] "PB(R6)"            "SPS(R)"            "dACA(R)"          
#> [124] "AL-DA1(R)"         "vACA(R)"           "PB(R7)"           
#> [127] "FBl7"              "AL-VM3(R)"         "LAL(-GA)(R)"      
#> [130] "SPS(L)"            "FBl1"              "AB(R)"            
#> [133] "hemibrain"         "PB(L1)"            "AL-DL4(L)"        
#> [136] "AL-DM2(L)"         "AL-VA1d(R)"        "SIP(L)"           
#> [139] "EB"                "AL-DA3(L)"         "PB(L6)"           
#> [142] "bL(L)"             "VES(L)"            "AL-VA3(R)"        
#> [145] "PENP"              "a2(R)"             "SNP(R)"           
#> [148] "AL-DA4m(R)"        "AL-VP1l(R)"        "NO1(R)"           
#> [151] "PB(R5)"            "ATL(L)"            "PB"               
#> [154] "NO3(L)"            "IB112"             "AL-DC4(R)"        
#> [157] "b'1(R)"            "PLP(R)"            "FBl6"             
#> [160] "VLNP(R)"           "mALT(L)"           "GC"               
#> [163] "b2(R)"             "SMP(L)"            "CRE(R)"           
#> [166] "IPS(R)"            "AL-VC3l(R)"        "POC"              
#> [169] "BU(R)"             "AL-DA4m(L)"        "AL(L)"            
#> [172] "PB(R3)"            "GF(R)"             "mALT(R)"          
#> [175] "aL(R)"             "NO3(R)"            "gL(L)"            
#> [178] "ICL(L)"            "AL-VC4(R)"         "AL-VM7v(R)"       
#> [181] "FBl2"              "b1(R)"             "EBr3am"           
#> [184] "ME(R)"             "b'L(L)"            "AL-DC2(R)"        
#> [187] "AL-DM4(R)"         "g2(R)"             "CA(R)"            
#> [190] "EPA(L)"            "AL-V(R)"           "AL-DL5(L)"        
#> [193] "LOP(R)"            "EBr3pw"            "FBl3"             
#> [196] "FBl8"              "AVLP(R)"           "aL(L)"            
#> [199] "a'L(L)"            "AL-DL1(R)"         "AL-VC1(R)"        
#> [202] "AL-VM1(R)"         "BU(L)"             "PB(L8)"           
#> [205] "ROB(R)"            "VES(R)"            "LX(L)"            
#> [208] "AL-DL5(R)"         "AL-VM5v(R)"        "MB(L)"            
#> [211] "CRE(L)"            "AL-VA2(R)"         "AL-VM7d(R)"       
#> [214] "VMNP"              "PB(R2)"            "AL-VP2(R)"        
#> [217] "NO2(L)"            "CA(L)"             "FBl5"             
#> [220] "AL-VA7m(R)"        "SCL(R)"            "AL-VC3m(R)"       
#> [223] "AL-VL2a(R)"        "SMP(R)"            "LX(R)"            
#> [226] "PB(R9)"            "PB(L9)"            "AL-DA3(R)"        
#> [229] "AL-DL3(R)"         "AL-VP1m(R)"        "g5(R)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(R1)"            "LAL(L)"            "FBl8"             
#>   [4] "a2(R)"             "FBl6"              "mALT(R)"          
#>   [7] "gL(L)"             "AB(L)"             "PRW"              
#>  [10] "ROB(R)"            "SAD(-AMMC)"        "PB(R3)"           
#>  [13] "SPS(L)"            "EBr1"              "EBr5"             
#>  [16] "a'L(R)"            "AL(L)"             "ATL(L)"           
#>  [19] "GF(R)"             "NO3(L)"            "a1(R)"            
#>  [22] "a'2(R)"            "dACA(R)"           "NO"               
#>  [25] "PB(L2)"            "NO3(R)"            "MB(L)"            
#>  [28] "FBl3"              "LX(L)"             "b'L(L)"           
#>  [31] "SCL(L)"            "PB(R7)"            "ATL(R)"           
#>  [34] "ICL(R)"            "LX(R)"             "PLP(R)"           
#>  [37] "ME(R)"             "bL(L)"             "FBl1"             
#>  [40] "FBl2"              "NO1(L)"            "b'L(R)"           
#>  [43] "SCL(R)"            "PB"                "PB(L6)"           
#>  [46] "IPS(R)"            "NO2(R)"            "a'L(L)"           
#>  [49] "LAL(-GA)(R)"       "EBr6"              "BU(L)"            
#>  [52] "NO(L)"             "a'3(R)"            "SIP(L)"           
#>  [55] "GA(R)"             "POC"               "PENP"             
#>  [58] "b2(R)"             "PB(L8)"            "EB"               
#>  [61] "MB(+ACA)(R)"       "RUB(L)"            "SPS(R)"           
#>  [64] "EPA(L)"            "MB(R)"             "PB(L9)"           
#>  [67] "EBr2r4"            "SIP(R)"            "CRE(L)"           
#>  [70] "LO(R)"             "VES(R)"            "FBl7"             
#>  [73] "bL(R)"             "aL(R)"             "vACA(R)"          
#>  [76] "GOR(R)"            "g5(R)"             "SNP(L)"           
#>  [79] "PVLP(R)"           "OL(R)"             "SMP(R)"           
#>  [82] "RUB(R)"            "SNP(R)"            "PB(L3)"           
#>  [85] "CRE(R)"            "PB(L7)"            "AME(R)"           
#>  [88] "PB(R4)"            "b'2(R)"            "EBr3am"           
#>  [91] "PB(L5)"            "CX"                "CA(L)"            
#>  [94] "IB"                "FBl5"              "aL(L)"            
#>  [97] "a3(R)"             "SMP(L)"            "PB(R8)"           
#> [100] "BU(R)"             "gL(R)"             "ICL(L)"           
#> [103] "LOP(R)"            "PB(R6)"            "EPA(R)"           
#> [106] "a'1(R)"            "b1(R)"             "AB(R)"            
#> [109] "WED(R)"            "PB(L1)"            "FLA(R)"           
#> [112] "VMNP"              "FBl4"              "AL(R)"            
#> [115] "hemibrain"         "g2(R)"             "NO(R)"            
#> [118] "PB(L4)"            "PED(R)"            "lACA(R)"          
#> [121] "FB"                "NO1(R)"            "PB(R9)"           
#> [124] "SAD"               "FBl9"              "PB(R2)"           
#> [127] "b'1(R)"            "EBr3d"             "CA(R)"            
#> [130] "LH(R)"             "CAN(R)"            "INP"              
#> [133] "VLNP(R)"           "NO2(L)"            "GC"               
#> [136] "AOTU(R)"           "g1(R)"             "SLP(R)"           
#> [139] "AMMC"              "g3(R)"             "AVLP(R)"          
#> [142] "GNG"               "VES(L)"            "AOT(R)"           
#> [145] "CRE(-ROB,-RUB)(R)" "g4(R)"             "EBr3pw"           
#> [148] "LAL(R)"            "PB(R5)"            "mALT(L)"          
#> [151] "GOR(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-VM3(R)"         "AL-V(R)"           "AL-VP5(R)"        
#>   [4] "AL-VP3(R)"         "GF(R)"             "ICL(R)"           
#>   [7] "AL-DM6(R)"         "b'L(L)"            "AL-DM3(R)"        
#>  [10] "IB"                "GC"                "VES(R)"           
#>  [13] "AL-DL1(R)"         "AL-DC4(R)"         "AL-VM7v(R)"       
#>  [16] "PB(L2)"            "AL-DM5(L)"         "FBl8"             
#>  [19] "AL-DA3(R)"         "EBr3pw"            "AL-VA1d(R)"       
#>  [22] "a2(R)"             "NO3(R)"            "EBr3d"            
#>  [25] "AL-VL2a(R)"        "ATL(L)"            "NO1(L)"           
#>  [28] "PB(L1)"            "AL-DA3(L)"         "MB(R)"            
#>  [31] "AL-VA6(R)"         "AL-VC5(R)"         "a1(R)"            
#>  [34] "AL(L)"             "AL-DL4(L)"         "PB(R4)"           
#>  [37] "dACA(R)"           "AL-DM2(L)"         "AL-VM5d(R)"       
#>  [40] "a'1(R)"            "EPA(L)"            "AL(R)"            
#>  [43] "g2(R)"             "EBr5"              "AL-DM4(L)"        
#>  [46] "FBl6"              "aL(L)"             "AL-DA4m(R)"       
#>  [49] "lACA(R)"           "MB(L)"             "CA(L)"            
#>  [52] "SPS(R)"            "PED(R)"            "mALT(R)"          
#>  [55] "PB(R7)"            "AL-VM7d(R)"        "CX"               
#>  [58] "mALT(L)"           "SMP(L)"            "ME(R)"            
#>  [61] "AL-VL2p(R)"        "FB"                "PB(L6)"           
#>  [64] "ICL(L)"            "SIP(L)"            "ROB(R)"           
#>  [67] "CRE(-RUB)(L)"      "SLP(R)"            "AL-VA5(R)"        
#>  [70] "AL-VL1(R)"         "PRW"               "FBl4"             
#>  [73] "LX(L)"             "AL-VC1(R)"         "LAL(L)"           
#>  [76] "LAL(R)"            "AL-DM1(R)"         "AL-VC3l(R)"       
#>  [79] "FB-column3"        "a3(R)"             "AOT(R)"           
#>  [82] "AL-VA3(R)"         "PLP(R)"            "AL-DC4(L)"        
#>  [85] "FBl9"              "AL-VP1d(R)"        "GA(R)"            
#>  [88] "PB(R1)"            "AL-DA2(R)"         "LAL(-GA)(R)"      
#>  [91] "WED(R)"            "LOP(R)"            "AL-DA4m(L)"       
#>  [94] "AL-DM3(L)"         "AL-VM1(R)"         "BU(R)"            
#>  [97] "IPS(R)"            "AL-VA1v(R)"        "SAD(-AMMC)"       
#> [100] "PB(R3)"            "AVLP(R)"           "AL-VP4(R)"        
#> [103] "g4(R)"             "AL-VM7v(L)"        "gL(R)"            
#> [106] "NO"                "AB(R)"             "AL-DL3(R)"        
#> [109] "AL-VA7l(R)"        "AL-VP1m(R)"        "GOR(L)"           
#> [112] "EBr1"              "EB"                "SMP(R)"           
#> [115] "NO2(R)"            "AL-DC1(R)"         "AL-VA4(R)"        
#> [118] "EBr6"              "EPA(R)"            "SNP(L)"           
#> [121] "b'1(R)"            "AMMC"              "BU(L)"            
#> [124] "PB(R8)"            "AL-DM6(L)"         "AL-DC3(R)"        
#> [127] "AL-VP2(R)"         "SCL(R)"            "gL(L)"            
#> [130] "RUB(L)"            "a'L(R)"            "AL-DC2(L)"        
#> [133] "FBl3"              "NO1(R)"            "LX(R)"            
#> [136] "NO(R)"             "RUB(R)"            "LH(R)"            
#> [139] "AL-D(R)"           "AL-VP1l(R)"        "AL-DL5(L)"        
#> [142] "CRE(R)"            "PB"                "FBl7"             
#> [145] "a'L(L)"            "PB(L5)"            "AOTU(R)"          
#> [148] "b'L(R)"            "CA(R)"             "AL-DA2(L)"        
#> [151] "CRE(L)"            "PB(L4)"            "AL-DM1(L)"        
#> [154] "SCL(L)"            "g3(R)"             "OL(R)"            
#> [157] "b2(R)"             "AL-DL5(R)"         "ATL(R)"           
#> [160] "AB(L)"             "GOR(R)"            "AL-VC2(R)"        
#> [163] "AL-DC1(L)"         "PB(L9)"            "GNG"              
#> [166] "AL-D(L)"           "AL-VA6(L)"         "AME(R)"           
#> [169] "vACA(R)"           "FLA(R)"            "PB(L8)"           
#> [172] "SIP(R)"            "NO(L)"             "AL-VM2(R)"        
#> [175] "AL-VM4(R)"         "a'3(R)"            "aL(R)"            
#> [178] "LO(R)"             "PVLP(R)"           "AL-DC2(R)"        
#> [181] "AL-DM5(R)"         "CRE(-ROB,-RUB)(R)" "a'2(R)"           
#> [184] "AL-DP1m(L)"        "AL-DA1(R)"         "PB(L3)"           
#> [187] "NO2(L)"            "PENP"              "MB(+ACA)(R)"      
#> [190] "AL-VA2(R)"         "PB(R6)"            "AL-DM2(R)"        
#> [193] "SAD"               "g1(R)"             "AL-DL2v(R)"       
#> [196] "VES(L)"            "VLNP(R)"           "SNP(R)"           
#> [199] "bL(L)"             "FBl1"              "NO3(L)"           
#> [202] "AL-DM4(R)"         "AL-DP1l(R)"        "AL-VC3m(R)"       
#> [205] "hemibrain"         "VMNP"              "AL-VM7d(L)"       
#> [208] "g5(R)"             "b1(R)"             "bL(R)"            
#> [211] "AL-VM5v(R)"        "CAN(R)"            "AL-DA4l(R)"       
#> [214] "AL-DP1m(R)"        "b'2(R)"            "PB(R2)"           
#> [217] "PB(L7)"            "EBr3am"            "PB(R5)"           
#> [220] "INP"               "AL-VC4(R)"         "POC"              
#> [223] "PB(R9)"            "AL-DL4(R)"         "SPS(L)"           
#> [226] "EBr2r4"            "AL-DL2d(R)"        "FBl2"             
#> [229] "FBl5"              "AL-VA7m(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "g3(R)"             "AL-VA6(L)"         "AL-VC1(R)"        
#>   [4] "AL-VP5(R)"         "AL-DM5(L)"         "LO(R)"            
#>   [7] "GNG"               "MB(R)"             "PB(L5)"           
#>  [10] "EBr1"              "IPS(R)"            "b'2(R)"           
#>  [13] "FLA(R)"            "AL-V(R)"           "bL(L)"            
#>  [16] "b'L(R)"            "AL-DP1m(L)"        "AL-DM5(R)"        
#>  [19] "AL-VA6(R)"         "AL-VM3(R)"         "AOT(R)"           
#>  [22] "AL-VM4(R)"         "vACA(R)"           "EPA(R)"           
#>  [25] "PB(L6)"            "AL-DC2(L)"         "POC"              
#>  [28] "LOP(R)"            "GOR(L)"            "g1(R)"            
#>  [31] "EPA(L)"            "AL-VL2p(R)"        "CRE(-ROB,-RUB)(R)"
#>  [34] "g2(R)"             "AL-DM2(L)"         "AL-DC1(L)"        
#>  [37] "AL-DM1(L)"         "VLNP(R)"           "PENP"             
#>  [40] "SPS(R)"            "aL(R)"             "SNP(L)"           
#>  [43] "PED(R)"            "AL-DM4(L)"         "NO(R)"            
#>  [46] "NO1(L)"            "FBl7"              "VMNP"             
#>  [49] "VES(L)"            "AL-DM4(R)"         "VES(R)"           
#>  [52] "a3(R)"             "AL-VC4(R)"         "AL-VP4(R)"        
#>  [55] "a'3(R)"            "SIP(R)"            "PB(L7)"           
#>  [58] "FBl2"              "AL-DL2v(R)"        "AL-DM3(R)"        
#>  [61] "AL-DP1l(R)"        "PB(R6)"            "LX(R)"            
#>  [64] "AL-DC3(R)"         "AL-VC2(R)"         "SCL(L)"           
#>  [67] "EBr3d"             "AL-DA3(L)"         "NO2(R)"           
#>  [70] "a'2(R)"            "ME(R)"             "PB(L9)"           
#>  [73] "FBl9"              "AL-VP1m(R)"        "ATL(R)"           
#>  [76] "INP"               "AL-VL2a(R)"        "SIP(L)"           
#>  [79] "NO"                "AL-DA4m(R)"        "AL-DC1(R)"        
#>  [82] "SMP(R)"            "AOTU(R)"           "GC"               
#>  [85] "EBr3pw"            "AME(R)"            "AL-VA7m(R)"       
#>  [88] "CA(R)"             "AL-DP1m(R)"        "AL-VA1d(R)"       
#>  [91] "mALT(L)"           "FBl4"              "AVLP(R)"          
#>  [94] "AL-VC5(R)"         "MB(L)"             "EBr2r4"           
#>  [97] "PB(R4)"            "OL(R)"             "AL-DA3(R)"        
#> [100] "AL-VA4(R)"         "PB(R2)"            "LAL(R)"           
#> [103] "SAD"               "ICL(L)"            "aL(L)"            
#> [106] "AL-DL5(R)"         "AL-VL1(R)"         "CRE(R)"           
#> [109] "SCL(R)"            "AL-DC4(L)"         "CA(L)"            
#> [112] "SPS(L)"            "ROB(R)"            "bL(R)"            
#> [115] "AL(R)"             "AL-DC2(R)"         "AL-DL1(R)"        
#> [118] "LAL(-GA)(R)"       "PVLP(R)"           "AL-VM5v(R)"       
#> [121] "PB(R5)"            "b'1(R)"            "a1(R)"            
#> [124] "LAL(L)"            "a2(R)"             "PB"               
#> [127] "a'L(R)"            "AL-VA7l(R)"        "AL-VM5d(R)"       
#> [130] "gL(L)"             "ATL(L)"            "EBr5"             
#> [133] "CRE(L)"            "AL-DA1(R)"         "AL-DM2(R)"        
#> [136] "SAD(-AMMC)"        "SLP(R)"            "AL-DM6(R)"        
#> [139] "AL-VA2(R)"         "AL-VA5(R)"         "PB(R8)"           
#> [142] "EB"                "AMMC"              "PB(R7)"           
#> [145] "b2(R)"             "NO3(L)"            "RUB(L)"           
#> [148] "AL-VA1v(R)"        "AL-VP1l(R)"        "AL(L)"            
#> [151] "PB(L4)"            "hemibrain"         "dACA(R)"          
#> [154] "AL-DA4m(L)"        "AL-DM3(L)"         "CRE(-RUB)(L)"     
#> [157] "g4(R)"             "AL-VM7d(L)"        "g5(R)"            
#> [160] "a'L(L)"            "AL-VA3(R)"         "AL-DL5(L)"        
#> [163] "AL-D(L)"           "CX"                "PRW"              
#> [166] "BU(R)"             "AL-VM7v(R)"        "NO3(R)"           
#> [169] "PB(L2)"            "FB"                "AL-DM6(L)"        
#> [172] "AL-DL2d(R)"        "AL-DM1(R)"         "AL-D(R)"          
#> [175] "AL-VM7d(R)"        "PB(L3)"            "LH(R)"            
#> [178] "PB(R3)"            "ICL(R)"            "AL-VC3l(R)"       
#> [181] "AL-VP1d(R)"        "FBl1"              "NO1(R)"           
#> [184] "PB(R1)"            "CAN(R)"            "AL-DL4(L)"        
#> [187] "gL(R)"             "FBl5"              "AL-DL3(R)"        
#> [190] "EBr3am"            "lACA(R)"           "b'L(L)"           
#> [193] "SMP(L)"            "PLP(R)"            "AL-VM7v(L)"       
#> [196] "AL-DC4(R)"         "mALT(R)"           "BU(L)"            
#> [199] "FBl6"              "AL-VP2(R)"         "AL-VM2(R)"        
#> [202] "GF(R)"             "AL-DA2(L)"         "PB(R9)"           
#> [205] "AL-VC3m(R)"        "IB"                "EBr6"             
#> [208] "FBl8"              "b1(R)"             "LX(L)"            
#> [211] "NO(L)"             "AL-DL4(R)"         "GA(R)"            
#> [214] "AB(L)"             "WED(R)"            "SNP(R)"           
#> [217] "PB(L8)"            "RUB(R)"            "PB(L1)"           
#> [220] "MB(+ACA)(R)"       "FBl3"              "AL-DA2(R)"        
#> [223] "AL-VM1(R)"         "AL-VP3(R)"         "a'1(R)"           
#> [226] "FB-column3"        "AB(R)"             "GOR(R)"           
#> [229] "AL-DA4l(R)"        "NO2(L)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method     path                                    
#> [1,] "REPORT"   "/api"                                  
#> [2,] "OPTIONS"  "/api"                                  
#> [3,] "PROPFIND" "/api"                                  
#> [4,] "POST"     "/api/raw/cypher/transaction/:id/cypher"
#> [5,] "POST"     "/api/v:ver/custom/custom"              
#> [6,] "POST"     "/api/raw/cypher/transaction"           
#>      name                                                                    
#> [1,] "github.com/labstack/echo.glob..func1"                                  
#> [2,] "github.com/labstack/echo.glob..func1"                                  
#> [3,] "github.com/labstack/echo.glob..func1"                                  
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    
#> [6,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }