Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-VM7v(L)"        "AL-DL4(R)"         "AOT(R)"           
#>   [4] "ICL(L)"            "CA(R)"             "IPS(R)"           
#>   [7] "b'L(L)"            "AL-DM2(R)"         "AL-VM7v(R)"       
#>  [10] "PB(R5)"            "AL-DL4(L)"         "AL-VM7d(L)"       
#>  [13] "PB(R1)"            "PB(L9)"            "PB(R4)"           
#>  [16] "FLA(R)"            "NO3(L)"            "FBl3"             
#>  [19] "GNG"               "AL-DA2(R)"         "GF(R)"            
#>  [22] "AL-DP1m(L)"        "FBl1"              "AL-DA4m(R)"       
#>  [25] "AL-VL1(R)"         "SAD"               "SNP(R)"           
#>  [28] "EBr3d"             "FBl2"              "AL-DL1(R)"        
#>  [31] "AL-VC5(R)"         "PRW"               "NO(R)"            
#>  [34] "AB(R)"             "AL-VC1(R)"         "EBr3am"           
#>  [37] "PB(L3)"            "AL-VA6(L)"         "FB-column3"       
#>  [40] "ME(R)"             "NO2(L)"            "AL-DM6(L)"        
#>  [43] "VES(L)"            "AL-DM3(R)"         "AL-VC4(R)"        
#>  [46] "SIP(L)"            "b2(R)"             "EB"               
#>  [49] "a3(R)"             "AL-DM1(R)"         "ICL(R)"           
#>  [52] "AL-VM2(R)"         "LAL(L)"            "g3(R)"            
#>  [55] "LAL(-GA)(R)"       "NO1(L)"            "AL-DM3(L)"        
#>  [58] "SNP(L)"            "RUB(L)"            "AL-VA5(R)"        
#>  [61] "WED(R)"            "CAN(R)"            "bL(R)"            
#>  [64] "a'L(L)"            "a'1(R)"            "SMP(R)"           
#>  [67] "vACA(R)"           "a1(R)"             "LH(R)"            
#>  [70] "GOR(R)"            "AL-DC4(R)"         "EBr5"             
#>  [73] "VMNP"              "AL-DA2(L)"         "ROB(R)"           
#>  [76] "PB(R3)"            "MB(+ACA)(R)"       "AL-DP1l(R)"       
#>  [79] "gL(L)"             "LX(L)"             "AL-VP2(R)"        
#>  [82] "AL-DA3(R)"         "EBr1"              "dACA(R)"          
#>  [85] "PB(L8)"            "AL-DM2(L)"         "AL-VP1d(R)"       
#>  [88] "a2(R)"             "EBr6"              "PB(L1)"           
#>  [91] "a'L(R)"            "PB(L2)"            "g2(R)"            
#>  [94] "aL(L)"             "AMMC"              "VES(R)"           
#>  [97] "AL-VM1(R)"         "AL-VA4(R)"         "CA(L)"            
#> [100] "AL-VM7d(R)"        "IB112"             "AL-DC2(R)"        
#> [103] "AL-DC2(L)"         "SIP(R)"            "PB(L6)"           
#> [106] "a'3(R)"            "SLP(R)"            "AL-DM6(R)"        
#> [109] "AME(R)"            "AL-VC2(R)"         "PB(L5)"           
#> [112] "VLNP(R)"           "b'L(R)"            "AL-DA4m(L)"       
#> [115] "AL-VC3m(R)"        "AL-VP5(R)"         "AL-VP1l(R)"       
#> [118] "PB(R7)"            "SCL(R)"            "AL-DM5(L)"        
#> [121] "b1(R)"             "AB(L)"             "PLP(R)"           
#> [124] "AL(R)"             "AL-VL2a(R)"        "lACA(R)"          
#> [127] "SPS(L)"            "AL-DL2d(R)"        "POC"              
#> [130] "g1(R)"             "NO3(R)"            "FBl6"             
#> [133] "AL-VA6(R)"         "AL-VC3l(R)"        "AL-D(R)"          
#> [136] "AL-VA2(R)"         "EBr2r4"            "g4(R)"            
#> [139] "NO(L)"             "AL-DC1(R)"         "AL-DP1m(R)"       
#> [142] "PB"                "FBl5"              "AL-DA4l(R)"       
#> [145] "AL-VA3(R)"         "SCL(L)"            "EBr3pw"           
#> [148] "AL-DL5(R)"         "FBl8"              "AL-VA1d(R)"       
#> [151] "b'2(R)"            "CRE(-RUB)(L)"      "SAD(-AMMC)"       
#> [154] "IB"                "SPS(R)"            "AL-VL2p(R)"       
#> [157] "CX"                "AL-DM1(L)"         "NO1(R)"           
#> [160] "CRE(L)"            "g5(R)"             "INP"              
#> [163] "mALT(L)"           "bL(L)"             "LO(R)"            
#> [166] "AL-VA7l(R)"        "LOP(R)"            "EPA(R)"           
#> [169] "OL(R)"             "AL-DA1(R)"         "AL-VP3(R)"        
#> [172] "PB(R8)"            "AL-DM4(L)"         "PB(L7)"           
#> [175] "EPA(L)"            "PB(L4)"            "BU(L)"            
#> [178] "NO"                "PVLP(R)"           "AL-VM5d(R)"       
#> [181] "PB(R6)"            "AL-DC3(R)"         "AL-DM4(R)"        
#> [184] "FBl9"              "AL-VM5v(R)"        "AL-VA7m(R)"       
#> [187] "PED(R)"            "AL-DL5(L)"         "AVLP(R)"          
#> [190] "AL-DL3(R)"         "AL-DM5(R)"         "NO2(R)"           
#> [193] "mALT(R)"           "MB(L)"             "gL(R)"            
#> [196] "PB(R2)"            "ATL(R)"            "PENP"             
#> [199] "AL-DC4(L)"         "LX(R)"             "BU(R)"            
#> [202] "AL-VP4(R)"         "RUB(R)"            "MB(R)"            
#> [205] "AL-DL2v(R)"        "AL-V(R)"           "hemibrain"        
#> [208] "CRE(R)"            "a'2(R)"            "AL-DC1(L)"        
#> [211] "AL-D(L)"           "AL-VP1m(R)"        "CRE(-ROB,-RUB)(R)"
#> [214] "GC"                "LAL(R)"            "AL-VM4(R)"        
#> [217] "ATL(L)"            "FB"                "AL(L)"            
#> [220] "AL-DA3(L)"         "PB(R9)"            "AL-VM3(R)"        
#> [223] "AL-VA1v(R)"        "b'1(R)"            "AOTU(R)"          
#> [226] "GA(R)"             "GOR(L)"            "aL(R)"            
#> [229] "FBl4"              "SMP(L)"            "FBl7"             
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB"                "VLNP(R)"           "PB(R5)"           
#>   [4] "NO2(L)"            "aL(R)"             "g2(R)"            
#>   [7] "PB(R8)"            "MB(+ACA)(R)"       "GA(R)"            
#>  [10] "MB(L)"             "EBr1"              "ME(R)"            
#>  [13] "AL(L)"             "SNP(L)"            "AOT(R)"           
#>  [16] "PB(L8)"            "gL(L)"             "PB(L6)"           
#>  [19] "AB(L)"             "PRW"               "b2(R)"            
#>  [22] "b'L(L)"            "a'3(R)"            "PB(L5)"           
#>  [25] "FBl1"              "aL(L)"             "PB(L4)"           
#>  [28] "mALT(L)"           "a1(R)"             "gL(R)"            
#>  [31] "FBl8"              "AL(R)"             "AOTU(R)"          
#>  [34] "NO(R)"             "SMP(R)"            "a'L(L)"           
#>  [37] "CA(L)"             "PB(R7)"            "FLA(R)"           
#>  [40] "EBr3pw"            "EBr2r4"            "PB(R3)"           
#>  [43] "bL(R)"             "IPS(R)"            "CAN(R)"           
#>  [46] "INP"               "g1(R)"             "EB"               
#>  [49] "g4(R)"             "SIP(L)"            "NO3(R)"           
#>  [52] "CRE(R)"            "SNP(R)"            "SIP(R)"           
#>  [55] "b'2(R)"            "GF(R)"             "PB(L3)"           
#>  [58] "EBr3am"            "ROB(R)"            "a'1(R)"           
#>  [61] "FBl3"              "ICL(R)"            "a3(R)"            
#>  [64] "MB(R)"             "EPA(L)"            "mALT(R)"          
#>  [67] "EBr5"              "AME(R)"            "FBl4"             
#>  [70] "OL(R)"             "RUB(R)"            "ATL(L)"           
#>  [73] "EBr6"              "PB(L1)"            "VES(R)"           
#>  [76] "EBr3d"             "RUB(L)"            "b'1(R)"           
#>  [79] "PENP"              "SMP(L)"            "SAD"              
#>  [82] "POC"               "LH(R)"             "LOP(R)"           
#>  [85] "PED(R)"            "EPA(R)"            "PVLP(R)"          
#>  [88] "LAL(R)"            "AMMC"              "VMNP"             
#>  [91] "PB(R1)"            "NO(L)"             "b1(R)"            
#>  [94] "NO1(R)"            "a'L(R)"            "CA(R)"            
#>  [97] "LX(L)"             "g3(R)"             "a2(R)"            
#> [100] "BU(L)"             "PB(L9)"            "g5(R)"            
#> [103] "CRE(-ROB,-RUB)(R)" "IB"                "LAL(-GA)(R)"      
#> [106] "NO"                "PB(R9)"            "SPS(L)"           
#> [109] "PLP(R)"            "BU(R)"             "FBl5"             
#> [112] "AB(R)"             "LAL(L)"            "FB"               
#> [115] "b'L(R)"            "SLP(R)"            "GOR(L)"           
#> [118] "ICL(L)"            "GOR(R)"            "SCL(L)"           
#> [121] "CX"                "GC"                "PB(R2)"           
#> [124] "VES(L)"            "FBl2"              "ATL(R)"           
#> [127] "GNG"               "WED(R)"            "dACA(R)"          
#> [130] "NO3(L)"            "PB(L2)"            "vACA(R)"          
#> [133] "NO1(L)"            "bL(L)"             "FBl6"             
#> [136] "PB(L7)"            "LO(R)"             "NO2(R)"           
#> [139] "PB(R6)"            "lACA(R)"           "a'2(R)"           
#> [142] "SAD(-AMMC)"        "FBl7"              "FBl9"             
#> [145] "SPS(R)"            "SCL(R)"            "CRE(L)"           
#> [148] "PB(R4)"            "hemibrain"         "LX(R)"            
#> [151] "AVLP(R)"          
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-VA6(R)"         "GF(R)"             "BU(R)"            
#>   [4] "AL-DM2(L)"         "AL-VM7d(L)"        "PB(R4)"           
#>   [7] "SAD"               "VES(R)"            "AL-VL2p(R)"       
#>  [10] "mALT(L)"           "NO(R)"             "SNP(L)"           
#>  [13] "FB-column3"        "PB(R2)"            "VES(L)"           
#>  [16] "FBl5"              "FBl7"              "AL-DA4l(R)"       
#>  [19] "AL-DP1m(R)"        "a'2(R)"            "IPS(R)"           
#>  [22] "lACA(R)"           "EPA(L)"            "AL-VC2(R)"        
#>  [25] "AL-DL4(R)"         "AL-VP4(R)"         "PB(L1)"           
#>  [28] "FBl8"              "LH(R)"             "LAL(L)"           
#>  [31] "LAL(-GA)(R)"       "AL(L)"             "AL-DC2(L)"        
#>  [34] "PVLP(R)"           "PB(L6)"            "b1(R)"            
#>  [37] "AL-DM5(R)"         "AL-DM6(R)"         "PENP"             
#>  [40] "SPS(R)"            "PED(R)"            "b'L(R)"           
#>  [43] "gL(L)"             "PB(R1)"            "GNG"              
#>  [46] "AL-VC4(R)"         "aL(R)"             "EBr5"             
#>  [49] "LO(R)"             "SLP(R)"            "AL-VA2(R)"        
#>  [52] "LOP(R)"            "LAL(R)"            "AL-VA7m(R)"       
#>  [55] "AL-VC3l(R)"        "AL-VC5(R)"         "AL-VM5d(R)"       
#>  [58] "GC"                "bL(R)"             "AL-VA3(R)"        
#>  [61] "AL-DM3(L)"         "PB(L2)"            "GA(R)"            
#>  [64] "AL-DL1(R)"         "AL-DL3(R)"         "EBr3am"           
#>  [67] "ATL(R)"            "AL-DA3(R)"         "AL-D(L)"          
#>  [70] "MB(R)"             "AL-VA1d(R)"        "PB(L5)"           
#>  [73] "ATL(L)"            "ICL(L)"            "AL-DL2d(R)"       
#>  [76] "AL-DM4(R)"         "SMP(L)"            "g3(R)"            
#>  [79] "NO3(R)"            "gL(R)"             "AL-DM3(R)"        
#>  [82] "PB(R3)"            "AL-DC3(R)"         "PB"               
#>  [85] "PLP(R)"            "AL-DC2(R)"         "CRE(-ROB,-RUB)(R)"
#>  [88] "AL-VA1v(R)"        "AL-VA4(R)"         "dACA(R)"          
#>  [91] "AL-DC4(L)"         "SIP(R)"            "a3(R)"            
#>  [94] "AL-DL2v(R)"        "AL(R)"             "IB"               
#>  [97] "CA(L)"             "a'1(R)"            "bL(L)"            
#> [100] "AL-DC4(R)"         "AL-DM2(R)"         "PB(R5)"           
#> [103] "CRE(R)"            "AL-DL5(L)"         "AL-DL4(L)"        
#> [106] "FBl6"              "aL(L)"             "AL-VP1l(R)"       
#> [109] "MB(L)"             "SNP(R)"            "SCL(R)"           
#> [112] "PB(R9)"            "b'2(R)"            "CRE(-RUB)(L)"     
#> [115] "AL-DP1l(R)"        "AL-VM4(R)"         "PB(L8)"           
#> [118] "AL-DA4m(L)"        "AL-DP1m(L)"        "PB(L9)"           
#> [121] "GOR(R)"            "AL-DL5(R)"         "PB(L3)"           
#> [124] "EBr6"              "FLA(R)"            "a'L(R)"           
#> [127] "AL-DM6(L)"         "FB"                "AL-VP5(R)"        
#> [130] "PB(R7)"            "AL-VA6(L)"         "AL-DM1(R)"        
#> [133] "NO2(L)"            "NO1(R)"            "PB(L4)"           
#> [136] "AL-VC1(R)"         "AMMC"              "AL-VM2(R)"        
#> [139] "AOTU(R)"           "EPA(R)"            "MB(+ACA)(R)"      
#> [142] "AL-DA1(R)"         "AL-VM1(R)"         "AL-DA4m(R)"       
#> [145] "g1(R)"             "BU(L)"             "AL-DM1(L)"        
#> [148] "FBl4"              "CAN(R)"            "SPS(L)"           
#> [151] "AL-DM5(L)"         "FBl1"              "CA(R)"            
#> [154] "ICL(R)"            "a2(R)"             "WED(R)"           
#> [157] "EB"                "AVLP(R)"           "AL-DC1(L)"        
#> [160] "LX(L)"             "AL-VA5(R)"         "AL-VC3m(R)"       
#> [163] "POC"               "OL(R)"             "AL-VP1m(R)"       
#> [166] "SIP(L)"            "AME(R)"            "PRW"              
#> [169] "PB(L7)"            "AOT(R)"            "PB(R6)"           
#> [172] "g2(R)"             "CRE(L)"            "a'L(L)"           
#> [175] "AL-D(R)"           "AL-V(R)"           "AL-VP3(R)"        
#> [178] "NO1(L)"            "AL-DA2(L)"         "FBl3"             
#> [181] "AL-VP2(R)"         "AL-VM7d(R)"        "LX(R)"            
#> [184] "AL-DA3(L)"         "AL-DM4(L)"         "NO"               
#> [187] "AL-VM3(R)"         "RUB(L)"            "AL-VA7l(R)"       
#> [190] "vACA(R)"           "a1(R)"             "AL-DC1(R)"        
#> [193] "AL-VM5v(R)"        "SMP(R)"            "RUB(R)"           
#> [196] "AL-VM7v(L)"        "AB(L)"             "EBr2r4"           
#> [199] "g4(R)"             "ROB(R)"            "NO2(R)"           
#> [202] "AL-VL1(R)"         "NO(L)"             "b'1(R)"           
#> [205] "hemibrain"         "g5(R)"             "FBl2"             
#> [208] "INP"               "AL-DA2(R)"         "GOR(L)"           
#> [211] "ME(R)"             "NO3(L)"            "EBr3pw"           
#> [214] "FBl9"              "CX"                "b2(R)"            
#> [217] "PB(R8)"            "EBr3d"             "AL-VP1d(R)"       
#> [220] "VLNP(R)"           "VMNP"              "b'L(L)"           
#> [223] "mALT(R)"           "EBr1"              "SAD(-AMMC)"       
#> [226] "AB(R)"             "SCL(L)"            "AL-VM7v(R)"       
#> [229] "AL-VL2a(R)"        "a'3(R)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-VC2(R)"         "RUB(R)"            "PB(L9)"           
#>   [4] "SNP(L)"            "PVLP(R)"           "LH(R)"            
#>   [7] "AL-VM4(R)"         "NO3(R)"            "PB(R1)"           
#>  [10] "AL(R)"             "AL-VP2(R)"         "GC"               
#>  [13] "PB(R9)"            "AVLP(R)"           "a'L(L)"           
#>  [16] "AL-VA4(R)"         "AL-DL5(L)"         "AL-DL2v(R)"       
#>  [19] "FLA(R)"            "a'2(R)"            "AL-DM4(L)"        
#>  [22] "AL-DC1(R)"         "AL-DC2(R)"         "AL-VA1d(R)"       
#>  [25] "PLP(R)"            "aL(L)"             "AL-DL5(R)"        
#>  [28] "AL-DM6(R)"         "GOR(L)"            "AL(L)"            
#>  [31] "AL-DP1m(L)"        "AOT(R)"            "AL-DC4(L)"        
#>  [34] "IPS(R)"            "SPS(R)"            "NO3(L)"           
#>  [37] "FBl8"              "b1(R)"             "LX(L)"            
#>  [40] "AL-DA4m(R)"        "AL-DL2d(R)"        "NO"               
#>  [43] "bL(R)"             "SMP(L)"            "SIP(L)"           
#>  [46] "FBl5"              "FBl6"              "CAN(R)"           
#>  [49] "SCL(L)"            "AL-DA3(L)"         "PB(R2)"           
#>  [52] "FB"                "EBr2r4"            "AB(R)"            
#>  [55] "AL-VL2a(R)"        "WED(R)"            "AL-VC3m(R)"       
#>  [58] "AL-VM3(R)"         "GF(R)"             "BU(L)"            
#>  [61] "AL-VM7v(R)"        "SPS(L)"            "g4(R)"            
#>  [64] "AL-DM5(L)"         "FBl4"              "AL-VP4(R)"        
#>  [67] "a'3(R)"            "AL-DM3(L)"         "AL-DL4(R)"        
#>  [70] "AL-VM5d(R)"        "NO1(L)"            "EBr5"             
#>  [73] "AL-DM2(L)"         "NO(R)"             "VES(R)"           
#>  [76] "ATL(L)"            "NO2(L)"            "PB(L3)"           
#>  [79] "GA(R)"             "AL-DA2(L)"         "EBr3pw"           
#>  [82] "mALT(L)"           "AL-VM7v(L)"        "INP"              
#>  [85] "PB(R6)"            "LX(R)"             "FB-column3"       
#>  [88] "AL-VL2p(R)"        "AL-VA6(L)"         "PB(L7)"           
#>  [91] "g3(R)"             "PB(R7)"            "BU(R)"            
#>  [94] "NO(L)"             "b'2(R)"            "LAL(L)"           
#>  [97] "PB"                "FBl1"              "AME(R)"           
#> [100] "GNG"               "AL-VP1m(R)"        "PB(L5)"           
#> [103] "PENP"              "lACA(R)"           "VES(L)"           
#> [106] "AL-VA7m(R)"        "AL-DC1(L)"         "SAD(-AMMC)"       
#> [109] "LO(R)"             "AL-DA4l(R)"        "CRE(-ROB,-RUB)(R)"
#> [112] "AL-VP1d(R)"        "NO2(R)"            "CA(L)"            
#> [115] "FBl3"              "AL-V(R)"           "b'1(R)"           
#> [118] "CX"                "AL-DL4(L)"         "AL-VM7d(L)"       
#> [121] "ATL(R)"            "AL-VC3l(R)"        "AL-VL1(R)"        
#> [124] "EBr1"              "SIP(R)"            "a3(R)"            
#> [127] "b'L(R)"            "PB(R3)"            "PB(R4)"           
#> [130] "AL-VC5(R)"         "AL-VP1l(R)"        "hemibrain"        
#> [133] "EPA(L)"            "AL-DA3(R)"         "EB"               
#> [136] "AL-DA4m(L)"        "GOR(R)"            "ICL(L)"           
#> [139] "PB(L4)"            "EBr3am"            "a2(R)"            
#> [142] "SAD"               "a'L(R)"            "AL-DM1(L)"        
#> [145] "AL-D(R)"           "vACA(R)"           "AL-DM6(L)"        
#> [148] "PB(L6)"            "AL-VA1v(R)"        "PB(L8)"           
#> [151] "AL-DM5(R)"         "AL-VM2(R)"         "VMNP"             
#> [154] "AL-VA7l(R)"        "AL-VC1(R)"         "AMMC"             
#> [157] "OL(R)"             "AL-VP3(R)"         "SCL(R)"           
#> [160] "SNP(R)"            "MB(+ACA)(R)"       "AL-DM2(R)"        
#> [163] "AL-VA3(R)"         "LAL(-GA)(R)"       "AL-VA2(R)"        
#> [166] "PB(R5)"            "AB(L)"             "CRE(R)"           
#> [169] "CRE(L)"            "ICL(R)"            "AL-DM1(R)"        
#> [172] "VLNP(R)"           "PED(R)"            "EBr6"             
#> [175] "EBr3d"             "AL-VA5(R)"         "g1(R)"            
#> [178] "dACA(R)"           "AL-DP1l(R)"        "a'1(R)"           
#> [181] "CA(R)"             "PB(L1)"            "AL-VM1(R)"        
#> [184] "AL-DC3(R)"         "AL-DC4(R)"         "PB(L2)"           
#> [187] "PRW"               "EPA(R)"            "AL-D(L)"          
#> [190] "FBl9"              "AL-VC4(R)"         "CRE(-RUB)(L)"     
#> [193] "b'L(L)"            "LAL(R)"            "AL-DA1(R)"        
#> [196] "AL-VA6(R)"         "aL(R)"             "MB(L)"            
#> [199] "AL-DC2(L)"         "SLP(R)"            "LOP(R)"           
#> [202] "g2(R)"             "FBl2"              "RUB(L)"           
#> [205] "AL-DL1(R)"         "AL-VM5v(R)"        "NO1(R)"           
#> [208] "ROB(R)"            "FBl7"              "AL-DM3(R)"        
#> [211] "AL-VM7d(R)"        "SMP(R)"            "PB(R8)"           
#> [214] "ME(R)"             "AL-VP5(R)"         "mALT(R)"          
#> [217] "gL(L)"             "AL-DM4(R)"         "b2(R)"            
#> [220] "g5(R)"             "AL-DL3(R)"         "POC"              
#> [223] "AOTU(R)"           "gL(R)"             "bL(L)"            
#> [226] "AL-DP1m(R)"        "IB"                "MB(R)"            
#> [229] "AL-DA2(R)"         "a1(R)"            
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method     path                                    
#> [1,] "POST"     "/api/npexplorer/neuronmeta"            
#> [2,] "GET"      "/api/roimeshes/mesh/:dataset/:roi"     
#> [3,] "POST"     "/api/raw/cypher/transaction/:id/cypher"
#> [4,] "GET"      "/token"                                
#> [5,] "GET"      "/api/npexplorer/autapses"              
#> [6,] "PROPFIND" "/api"                                  
#>      name                                                                                      
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMeta-fm"
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/roimeshes.masterAPI.getMesh-fm"          
#> [3,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                  
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                  
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"  
#> [6,] "github.com/labstack/echo.glob..func1"                                                    

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }