Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "SPS(L)"            "a'1(R)"            "GNG"              
#>   [4] "PB(R5)"            "PB(L2)"            "AMMC"             
#>   [7] "FB"                "EBr6"              "PB(L7)"           
#>  [10] "MB(R)"             "POC"               "EBr3am"           
#>  [13] "CX"                "LAL(L)"            "dACA(R)"          
#>  [16] "PB(L8)"            "EBr2r4"            "PB(R2)"           
#>  [19] "b'L(L)"            "hemibrain"         "GC"               
#>  [22] "FBl7"              "CA(L)"             "PB(L3)"           
#>  [25] "LAL(-GA)(R)"       "WED(R)"            "NO1(L)"           
#>  [28] "a'2(R)"            "PB(R9)"            "SMP(L)"           
#>  [31] "PLP(R)"            "LO(R)"             "RUB(R)"           
#>  [34] "CRE(-ROB,-RUB)(R)" "SMP(R)"            "b'L(R)"           
#>  [37] "BU(L)"             "gL(R)"             "NO3(L)"           
#>  [40] "g5(R)"             "PVLP(R)"           "g1(R)"            
#>  [43] "LX(R)"             "g4(R)"             "ICL(L)"           
#>  [46] "GF(R)"             "a'L(R)"            "EBr5"             
#>  [49] "BU(R)"             "b'1(R)"            "NO2(L)"           
#>  [52] "OL(R)"             "SLP(R)"            "GOR(L)"           
#>  [55] "PB(L1)"            "PB"                "FBl9"             
#>  [58] "PB(R4)"            "LAL(R)"            "gL(L)"            
#>  [61] "FBl3"              "FBl5"              "a'3(R)"           
#>  [64] "FBl6"              "b1(R)"             "AB(R)"            
#>  [67] "CAN(R)"            "ME(R)"             "EBr3d"            
#>  [70] "SNP(L)"            "FBl1"              "ATL(R)"           
#>  [73] "AOT(R)"            "GOR(R)"            "GA(R)"            
#>  [76] "SIP(L)"            "NO(R)"             "FBl2"             
#>  [79] "ICL(R)"            "LX(L)"             "PB(R8)"           
#>  [82] "AL(L)"             "SIP(R)"            "MB(L)"            
#>  [85] "aL(L)"             "VES(L)"            "mALT(R)"          
#>  [88] "a'L(L)"            "LOP(R)"            "g2(R)"            
#>  [91] "SCL(R)"            "ROB(R)"            "CA(R)"            
#>  [94] "PRW"               "PB(R7)"            "a3(R)"            
#>  [97] "AL(R)"             "vACA(R)"           "AVLP(R)"          
#> [100] "PB(L4)"            "PENP"              "SAD"              
#> [103] "EBr1"              "CRE(L)"            "FBl4"             
#> [106] "FBl8"              "PB(L5)"            "EPA(R)"           
#> [109] "AME(R)"            "VLNP(R)"           "ATL(L)"           
#> [112] "b'2(R)"            "NO1(R)"            "VMNP"             
#> [115] "LH(R)"             "INP"               "IB"               
#> [118] "PB(R6)"            "AOTU(R)"           "NO"               
#> [121] "PB(R3)"            "aL(R)"             "FLA(R)"           
#> [124] "SNP(R)"            "SAD(-AMMC)"        "MB(+ACA)(R)"      
#> [127] "NO2(R)"            "mALT(L)"           "g3(R)"            
#> [130] "CRE(R)"            "IPS(R)"            "bL(R)"            
#> [133] "NO(L)"             "AB(L)"             "SPS(R)"           
#> [136] "lACA(R)"           "SCL(L)"            "PED(R)"           
#> [139] "bL(L)"             "NO3(R)"            "PB(R1)"           
#> [142] "PB(L9)"            "EBr3pw"            "a2(R)"            
#> [145] "b2(R)"             "EB"                "PB(L6)"           
#> [148] "VES(R)"            "EPA(L)"            "RUB(L)"           
#> [151] "a1(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                                        
#> [1,] "CONNECT" "/api/*"                                    
#> [2,] "GET"     "/api/npexplorer/neuronmetavals"            
#> [3,] "GET"     "/api/v:ver/npexplorer/simpleconnections"   
#> [4,] "POST"    "/api/v:ver/skeletons/skeleton/:dataset/:id"
#> [5,] "GET"     "/api/v:ver/dbmeta/database"                
#> [6,] "GET"     "/login"                                    
#>      name                                                                                          
#> [1,] "github.com/labstack/echo.glob..func1"                                                        
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMetaVals-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
#> [6,] "github.com/janelia-flyem/echo-secure.loginHandler"                                           

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }