Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "FBl5"              "vACA(R)"           "SAD"              
#>   [4] "WED(R)"            "EBr5"              "PB(L1)"           
#>   [7] "PLP(R)"            "FBl1"              "a1(R)"            
#>  [10] "PB(L3)"            "PB(R7)"            "GC"               
#>  [13] "PED(R)"            "PB(R3)"            "RUB(L)"           
#>  [16] "SCL(L)"            "LOP(R)"            "b1(R)"            
#>  [19] "GOR(R)"            "MB(R)"             "PENP"             
#>  [22] "FBl7"              "FBl9"              "EBr3pw"           
#>  [25] "IPS(R)"            "CAN(R)"            "AOTU(R)"          
#>  [28] "NO3(R)"            "PB(R9)"            "CRE(R)"           
#>  [31] "FLA(R)"            "AB(R)"             "b'L(R)"           
#>  [34] "PB(R2)"            "a3(R)"             "g3(R)"            
#>  [37] "AL(L)"             "SNP(L)"            "EBr6"             
#>  [40] "PB(L7)"            "POC"               "PB(R1)"           
#>  [43] "AVLP(R)"           "LX(L)"             "b'L(L)"           
#>  [46] "mALT(R)"           "EBr1"              "NO1(R)"           
#>  [49] "g4(R)"             "bL(L)"             "NO2(R)"           
#>  [52] "PB(R4)"            "LAL(R)"            "b2(R)"            
#>  [55] "MB(L)"             "EBr3d"             "a'3(R)"           
#>  [58] "mALT(L)"           "b'1(R)"            "FBl4"             
#>  [61] "ICL(R)"            "AOT(R)"            "BU(L)"            
#>  [64] "SPS(L)"            "AL(R)"             "CRE(-ROB,-RUB)(R)"
#>  [67] "FBl8"              "SMP(R)"            "b'2(R)"           
#>  [70] "LAL(L)"            "g2(R)"             "GF(R)"            
#>  [73] "g1(R)"             "AME(R)"            "PB(L5)"           
#>  [76] "NO1(L)"            "SAD(-AMMC)"        "FBl2"             
#>  [79] "PB(L2)"            "a'1(R)"            "INP"              
#>  [82] "a'L(L)"            "RUB(R)"            "a2(R)"            
#>  [85] "CRE(L)"            "FBl3"              "aL(L)"            
#>  [88] "OL(R)"             "LAL(-GA)(R)"       "FBl6"             
#>  [91] "ATL(R)"            "gL(R)"             "g5(R)"            
#>  [94] "NO(L)"             "PB(R5)"            "AMMC"             
#>  [97] "SIP(L)"            "PB(R8)"            "a'2(R)"           
#> [100] "NO(R)"             "PRW"               "GOR(L)"           
#> [103] "dACA(R)"           "SPS(R)"            "lACA(R)"          
#> [106] "VES(L)"            "PVLP(R)"           "GNG"              
#> [109] "CX"                "SMP(L)"            "NO"               
#> [112] "ICL(L)"            "PB(L6)"            "SCL(R)"           
#> [115] "VMNP"              "gL(L)"             "CA(L)"            
#> [118] "NO2(L)"            "PB"                "ME(R)"            
#> [121] "VES(R)"            "SNP(R)"            "LH(R)"            
#> [124] "EPA(L)"            "GA(R)"             "hemibrain"        
#> [127] "LX(R)"             "LO(R)"             "bL(R)"            
#> [130] "ATL(L)"            "FB"                "a'L(R)"           
#> [133] "BU(R)"             "EBr3am"            "EBr2r4"           
#> [136] "EB"                "IB"                "ROB(R)"           
#> [139] "PB(L9)"            "PB(L4)"            "SLP(R)"           
#> [142] "AB(L)"             "PB(R6)"            "EPA(R)"           
#> [145] "SIP(R)"            "MB(+ACA)(R)"       "CA(R)"            
#> [148] "VLNP(R)"           "aL(R)"             "PB(L8)"           
#> [151] "NO3(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                              
#> [1,] "GET"     "/api/custom/custom"              
#> [2,] "POST"    "/api/v:ver/custom/custom"        
#> [3,] "GET"     "/api/v:ver/npexplorer/neuronmeta"
#> [4,] "GET"     "/api/npexplorer/roisinneuron"    
#> [5,] "POST"    "/logout"                         
#> [6,] "OPTIONS" "/api"                            
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"             
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"
#> [5,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [6,] "github.com/labstack/echo.glob..func1"                                                      

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }