Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "FBl9"              "VES(L)"            "AL-DA2(R)"        
#>   [4] "AL-DA4l(R)"        "AL-DM1(R)"         "PB(R5)"           
#>   [7] "PB(L2)"            "AL-DM3(R)"         "AL-VP1m(R)"       
#>  [10] "EPA(R)"            "FBl7"              "AL-DC1(R)"        
#>  [13] "AL-DM6(R)"         "AL-DP1l(R)"        "AL-VP2(R)"        
#>  [16] "PENP"              "GOR(L)"            "g1(R)"            
#>  [19] "AME(R)"            "bL(R)"             "FLA(R)"           
#>  [22] "a'2(R)"            "AL-DM4(L)"         "AL-VM7v(R)"       
#>  [25] "vACA(R)"           "WED(R)"            "AL-DC4(L)"        
#>  [28] "IPS(R)"            "IB"                "a'3(R)"           
#>  [31] "SCL(R)"            "BU(R)"             "AL-VM5d(R)"       
#>  [34] "AL-VP1l(R)"        "AMMC"              "a1(R)"            
#>  [37] "AOTU(R)"           "AL-VA1v(R)"        "AL-VL2p(R)"       
#>  [40] "NO"                "SMP(R)"            "GC"               
#>  [43] "PB"                "SIP(R)"            "PB(R9)"           
#>  [46] "AL-VP1d(R)"        "SAD"               "b'L(R)"           
#>  [49] "FBl4"              "a2(R)"             "EBr5"             
#>  [52] "PB(L8)"            "AL-DL4(L)"         "PB(R4)"           
#>  [55] "SPS(L)"            "AL-DM6(L)"         "b1(R)"            
#>  [58] "b'2(R)"            "EBr1"              "NO1(R)"           
#>  [61] "LX(R)"             "EBr3d"             "AL-DC2(R)"        
#>  [64] "AL-VA3(R)"         "AL-DA3(L)"         "FBl5"             
#>  [67] "PB(R3)"            "AL-DM4(R)"         "ATL(L)"           
#>  [70] "LAL(-GA)(R)"       "AL-DA4m(L)"        "CX"               
#>  [73] "NO3(R)"            "BU(L)"             "FBl8"             
#>  [76] "aL(L)"             "PB(L4)"            "AL-VA4(R)"        
#>  [79] "SNP(R)"            "a'1(R)"            "LX(L)"            
#>  [82] "PB(R7)"            "AL-DM2(L)"         "gL(L)"            
#>  [85] "bL(L)"             "INP"               "AL-VM5v(R)"       
#>  [88] "AL-V(R)"           "EBr3am"            "NO3(L)"           
#>  [91] "CRE(-RUB)(L)"      "EPA(L)"            "AL-DC1(L)"        
#>  [94] "PB(L6)"            "gL(R)"             "FBl3"             
#>  [97] "AL-VL1(R)"         "SAD(-AMMC)"        "NO2(L)"           
#> [100] "AL-DC2(L)"         "ROB(R)"            "FBl6"             
#> [103] "CA(R)"             "SMP(L)"            "AL(R)"            
#> [106] "AL-VC4(R)"         "AL-DL5(L)"         "RUB(R)"           
#> [109] "AL-VP4(R)"         "LAL(L)"            "EB"               
#> [112] "AL-VM7v(L)"        "EBr3pw"            "FB-column3"       
#> [115] "AL-DL3(R)"         "AL-VM4(R)"         "mALT(R)"          
#> [118] "VMNP"              "IB112"             "hemibrain"        
#> [121] "PLP(R)"            "g4(R)"             "GNG"              
#> [124] "b'L(L)"            "AL-VC1(R)"         "GA(R)"            
#> [127] "ATL(R)"            "AL-DA3(R)"         "AL-DA4m(R)"       
#> [130] "CRE(-ROB,-RUB)(R)" "PB(L3)"            "PRW"              
#> [133] "NO(R)"             "OL(R)"             "AL-VC3m(R)"       
#> [136] "AL-DL2d(R)"        "PB(L7)"            "MB(+ACA)(R)"      
#> [139] "SPS(R)"            "PB(R1)"            "a3(R)"            
#> [142] "AL-DA1(R)"         "AL-VC3l(R)"        "PB(R6)"           
#> [145] "AL-VP3(R)"         "AB(L)"             "AL-VP5(R)"        
#> [148] "EBr6"              "AL-DM3(L)"         "PVLP(R)"          
#> [151] "AL-DC3(R)"         "AL-DP1m(R)"        "AL-D(R)"          
#> [154] "AL-VC5(R)"         "FB"                "AL(L)"            
#> [157] "AL-VM7d(R)"        "NO1(L)"            "b2(R)"            
#> [160] "PB(L1)"            "PB(R2)"            "AVLP(R)"          
#> [163] "AL-VA6(R)"         "LO(R)"             "FBl1"             
#> [166] "AL-VM7d(L)"        "MB(R)"             "LOP(R)"           
#> [169] "CRE(R)"            "AB(R)"             "ME(R)"            
#> [172] "AL-VA6(L)"         "CAN(R)"            "NO(L)"            
#> [175] "a'L(L)"            "CA(L)"             "AL-DM1(L)"        
#> [178] "VES(R)"            "AL-VC2(R)"         "PB(L5)"           
#> [181] "g3(R)"             "mALT(L)"           "aL(R)"            
#> [184] "AL-DL4(R)"         "dACA(R)"           "AL-D(L)"          
#> [187] "NO2(R)"            "AL-DC4(R)"         "SCL(L)"           
#> [190] "AL-VA2(R)"         "AL-VA7l(R)"        "ICL(L)"           
#> [193] "g5(R)"             "SNP(L)"            "ICL(R)"           
#> [196] "GOR(R)"            "LAL(R)"            "SLP(R)"           
#> [199] "GF(R)"             "SIP(L)"            "EBr2r4"           
#> [202] "AL-DA2(L)"         "RUB(L)"            "AL-VA5(R)"        
#> [205] "MB(L)"             "AL-VA1d(R)"        "AL-DP1m(L)"       
#> [208] "PB(L9)"            "AL-DM5(R)"         "AL-VM2(R)"        
#> [211] "VLNP(R)"           "AL-DL5(R)"         "AL-VA7m(R)"       
#> [214] "lACA(R)"           "AL-DL2v(R)"        "POC"              
#> [217] "g2(R)"             "PED(R)"            "PB(R8)"           
#> [220] "FBl2"              "AL-DM2(R)"         "AOT(R)"           
#> [223] "LH(R)"             "a'L(R)"            "CRE(L)"           
#> [226] "AL-DL1(R)"         "AL-VL2a(R)"        "AL-VM1(R)"        
#> [229] "AL-VM3(R)"         "AL-DM5(L)"         "b'1(R)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "LH(R)"             "a'L(L)"            "BU(L)"            
#>   [4] "SPS(L)"            "PB(L6)"            "ICL(L)"           
#>   [7] "SAD(-AMMC)"        "VES(R)"            "PLP(R)"           
#>  [10] "a'2(R)"            "FBl5"              "PRW"              
#>  [13] "FB"                "EBr3pw"            "AL(R)"            
#>  [16] "NO3(R)"            "FBl9"              "OL(R)"            
#>  [19] "SCL(R)"            "dACA(R)"           "PB(L7)"           
#>  [22] "AB(R)"             "NO2(R)"            "PB(R5)"           
#>  [25] "RUB(R)"            "EBr3am"            "mALT(L)"          
#>  [28] "PB(R9)"            "ROB(R)"            "SNP(L)"           
#>  [31] "AVLP(R)"           "LAL(-GA)(R)"       "LX(R)"            
#>  [34] "a'L(R)"            "SIP(R)"            "NO(R)"            
#>  [37] "PB(L9)"            "b'1(R)"            "PB(L2)"           
#>  [40] "a'3(R)"            "bL(R)"             "NO(L)"            
#>  [43] "SAD"               "WED(R)"            "CRE(R)"           
#>  [46] "EBr6"              "PB"                "EBr2r4"           
#>  [49] "PB(R3)"            "AOT(R)"            "LOP(R)"           
#>  [52] "CX"                "SMP(R)"            "NO2(L)"           
#>  [55] "GA(R)"             "FBl1"              "FBl6"             
#>  [58] "b'L(L)"            "LAL(R)"            "VES(L)"           
#>  [61] "hemibrain"         "BU(R)"             "CA(R)"            
#>  [64] "NO"                "MB(+ACA)(R)"       "mALT(R)"          
#>  [67] "gL(L)"             "AME(R)"            "lACA(R)"          
#>  [70] "NO1(L)"            "bL(L)"             "EPA(L)"           
#>  [73] "PENP"              "SIP(L)"            "EBr5"             
#>  [76] "PB(L8)"            "AB(L)"             "g3(R)"            
#>  [79] "EBr3d"             "SNP(R)"            "b1(R)"            
#>  [82] "IPS(R)"            "b2(R)"             "VMNP"             
#>  [85] "ME(R)"             "g4(R)"             "CAN(R)"           
#>  [88] "a1(R)"             "PB(R2)"            "ATL(R)"           
#>  [91] "AOTU(R)"           "PED(R)"            "g2(R)"            
#>  [94] "gL(R)"             "LO(R)"             "FBl2"             
#>  [97] "aL(L)"             "INP"               "PB(R4)"           
#> [100] "PB(L5)"            "LAL(L)"            "SPS(R)"           
#> [103] "PB(R7)"            "CRE(L)"            "ICL(R)"           
#> [106] "FBl7"              "MB(R)"             "PB(L1)"           
#> [109] "GNG"               "vACA(R)"           "ATL(L)"           
#> [112] "AMMC"              "EB"                "b'2(R)"           
#> [115] "NO1(R)"            "EPA(R)"            "PB(L4)"           
#> [118] "SCL(L)"            "SMP(L)"            "g5(R)"            
#> [121] "FBl4"              "SLP(R)"            "PB(R6)"           
#> [124] "GC"                "g1(R)"             "PB(R8)"           
#> [127] "LX(L)"             "NO3(L)"            "IB"               
#> [130] "PB(L3)"            "aL(R)"             "FLA(R)"           
#> [133] "PVLP(R)"           "PB(R1)"            "a3(R)"            
#> [136] "GF(R)"             "a2(R)"             "AL(L)"            
#> [139] "CRE(-ROB,-RUB)(R)" "GOR(L)"            "b'L(R)"           
#> [142] "a'1(R)"            "RUB(L)"            "VLNP(R)"          
#> [145] "POC"               "MB(L)"             "EBr1"             
#> [148] "FBl3"              "FBl8"              "GOR(R)"           
#> [151] "CA(L)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "EBr3d"             "AL-DA3(L)"         "a'L(L)"           
#>   [4] "AL-VP1d(R)"        "AB(L)"             "GNG"              
#>   [7] "AL-VL1(R)"         "AL-VM7d(R)"        "PB(R3)"           
#>  [10] "AME(R)"            "bL(R)"             "RUB(R)"           
#>  [13] "BU(L)"             "NO"                "AL-DM2(L)"        
#>  [16] "AL-DA3(R)"         "AL-DM5(R)"         "LAL(-GA)(R)"      
#>  [19] "WED(R)"            "AL-DP1m(L)"        "AL-DC2(R)"        
#>  [22] "AL-DC3(R)"         "FB"                "dACA(R)"          
#>  [25] "PB(L9)"            "CAN(R)"            "AL-VA1d(R)"       
#>  [28] "AL-VA1v(R)"        "NO1(R)"            "SPS(L)"           
#>  [31] "ICL(L)"            "a'3(R)"            "a1(R)"            
#>  [34] "GF(R)"             "mALT(R)"           "b'L(R)"           
#>  [37] "NO(R)"             "PLP(R)"            "AVLP(R)"          
#>  [40] "SCL(L)"            "LAL(L)"            "GA(R)"            
#>  [43] "ROB(R)"            "AL-DP1l(R)"        "a'L(R)"           
#>  [46] "EPA(R)"            "AL-DM3(L)"         "IPS(R)"           
#>  [49] "AL-VA5(R)"         "PB(L2)"            "PB(L3)"           
#>  [52] "LOP(R)"            "AL-DM4(L)"         "FBl4"             
#>  [55] "CA(R)"             "ATL(R)"            "ICL(R)"           
#>  [58] "FLA(R)"            "AL-DA4m(R)"        "AL-DM3(R)"        
#>  [61] "AL-D(R)"           "VLNP(R)"           "EBr2r4"           
#>  [64] "AL-VM7d(L)"        "LX(L)"             "GOR(L)"           
#>  [67] "CRE(R)"            "bL(L)"             "OL(R)"            
#>  [70] "AL(R)"             "AL-VM7v(L)"        "AL-DA2(R)"        
#>  [73] "AL-VL2a(R)"        "AL-VL2p(R)"        "AL-VP5(R)"        
#>  [76] "CRE(-ROB,-RUB)(R)" "AL-VA7l(R)"        "AL-VC1(R)"        
#>  [79] "PED(R)"            "MB(L)"             "g2(R)"            
#>  [82] "AL-VA6(L)"         "AL-DA1(R)"         "AL-DC4(R)"        
#>  [85] "EPA(L)"            "PB(R4)"            "NO(L)"            
#>  [88] "AL-DC1(R)"         "POC"               "PB(R9)"           
#>  [91] "PB(L4)"            "a2(R)"             "AL-DM6(L)"        
#>  [94] "PB(L6)"            "AL-VP1l(R)"        "b'1(R)"           
#>  [97] "EBr6"              "LX(R)"             "AL-DA4m(L)"       
#> [100] "FBl2"              "AL-DM6(R)"         "AL-VM2(R)"        
#> [103] "hemibrain"         "lACA(R)"           "aL(R)"            
#> [106] "AL-DA2(L)"         "AL-DM4(R)"         "ME(R)"            
#> [109] "AL-VC3m(R)"        "PB(R6)"            "SMP(L)"           
#> [112] "AL-DC1(L)"         "CRE(-RUB)(L)"      "AL-VM7v(R)"       
#> [115] "EBr3am"            "PB(L5)"            "NO1(L)"           
#> [118] "MB(+ACA)(R)"       "AB(R)"             "b'L(L)"           
#> [121] "PB"                "g4(R)"             "SMP(R)"           
#> [124] "AL-DC2(L)"         "AL-DL4(L)"         "GOR(R)"           
#> [127] "AL-VC5(R)"         "gL(L)"             "AL-D(L)"          
#> [130] "CRE(L)"            "NO2(R)"            "AL-VM3(R)"        
#> [133] "ATL(L)"            "PB(L8)"            "LO(R)"            
#> [136] "g3(R)"             "AL-VA3(R)"         "AL-VC3l(R)"       
#> [139] "FBl5"              "AL-VM5v(R)"        "AL-DC4(L)"        
#> [142] "SIP(R)"            "PB(L7)"            "a3(R)"            
#> [145] "mALT(L)"           "AL-VM1(R)"         "AL-VP1m(R)"       
#> [148] "gL(R)"             "g5(R)"             "SNP(L)"           
#> [151] "FBl1"              "AL-VA6(R)"         "AL-VA7m(R)"       
#> [154] "g1(R)"             "AL-DM2(R)"         "AL-VP3(R)"        
#> [157] "SIP(L)"            "PB(R8)"            "aL(L)"            
#> [160] "IB"                "VMNP"              "SAD(-AMMC)"       
#> [163] "FBl3"              "AL-DM5(L)"         "b1(R)"            
#> [166] "AL-DL3(R)"         "AOT(R)"            "AMMC"             
#> [169] "NO2(L)"            "SCL(R)"            "VES(R)"           
#> [172] "CX"                "EBr5"              "BU(R)"            
#> [175] "AL-DM1(L)"         "PVLP(R)"           "AL-VA2(R)"        
#> [178] "PB(R5)"            "SPS(R)"            "INP"              
#> [181] "a'1(R)"            "AL-V(R)"           "RUB(L)"           
#> [184] "AL(L)"             "FB-column3"        "FBl9"             
#> [187] "AL-DL4(R)"         "AL-VM4(R)"         "FBl7"             
#> [190] "AL-DL1(R)"         "AL-VC2(R)"         "b'2(R)"           
#> [193] "PRW"               "AL-DL5(L)"         "LH(R)"            
#> [196] "AL-DA4l(R)"        "AL-VP4(R)"         "EBr3pw"           
#> [199] "FBl8"              "AL-DL5(R)"         "CA(L)"            
#> [202] "PB(R1)"            "PB(R2)"            "GC"               
#> [205] "b2(R)"             "NO3(L)"            "SLP(R)"           
#> [208] "MB(R)"             "AL-DL2d(R)"        "PENP"             
#> [211] "EBr1"              "AL-DL2v(R)"        "AL-DP1m(R)"       
#> [214] "AL-VP2(R)"         "vACA(R)"           "EB"               
#> [217] "AL-VM5d(R)"        "FBl6"              "AOTU(R)"          
#> [220] "PB(R7)"            "a'2(R)"            "VES(L)"           
#> [223] "AL-VC4(R)"         "SAD"               "NO3(R)"           
#> [226] "PB(L1)"            "LAL(R)"            "AL-VA4(R)"        
#> [229] "SNP(R)"            "AL-DM1(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "PB(L6)"            "AL-DL2v(R)"        "AL-VM5v(R)"       
#>   [4] "AB(L)"             "RUB(R)"            "AL-DC4(L)"        
#>   [7] "ATL(R)"            "AL-VC3m(R)"        "AL-VM7d(R)"       
#>  [10] "b1(R)"             "AL-VP3(R)"         "AL-VP1d(R)"       
#>  [13] "AVLP(R)"           "LX(L)"             "GNG"              
#>  [16] "bL(R)"             "g1(R)"             "FB"               
#>  [19] "g4(R)"             "AL(L)"             "AL-DM4(R)"        
#>  [22] "EBr1"              "CRE(L)"            "LAL(R)"           
#>  [25] "b'L(R)"            "FBl7"              "FBl8"             
#>  [28] "a3(R)"             "SIP(R)"            "AL-DC4(R)"        
#>  [31] "AL-D(R)"           "b2(R)"             "LH(R)"            
#>  [34] "AL-DP1l(R)"        "CRE(-ROB,-RUB)(R)" "gL(R)"            
#>  [37] "AL-VA7m(R)"        "a1(R)"             "SPS(L)"           
#>  [40] "PB(L4)"            "AL-DM1(R)"         "VMNP"             
#>  [43] "g5(R)"             "NO2(R)"            "AL-DL5(R)"        
#>  [46] "b'L(L)"            "VES(R)"            "MB(R)"            
#>  [49] "AL-DP1m(R)"        "AL-VM7v(R)"        "CAN(R)"           
#>  [52] "AL-VA4(R)"         "GA(R)"             "AL-VM5d(R)"       
#>  [55] "PB(R9)"            "FBl5"              "SPS(R)"           
#>  [58] "FBl1"              "AL-VP1m(R)"        "PB(R5)"           
#>  [61] "mALT(L)"           "LX(R)"             "AL-DA2(L)"        
#>  [64] "bL(L)"             "AL-DL4(R)"         "AL-VC5(R)"        
#>  [67] "PED(R)"            "FLA(R)"            "a'L(R)"           
#>  [70] "AL-VC3l(R)"        "ATL(L)"            "PENP"             
#>  [73] "PLP(R)"            "AL-DM6(L)"         "AL-DL2d(R)"       
#>  [76] "a2(R)"             "EBr2r4"            "IB"               
#>  [79] "PB(L5)"            "FBl2"              "EBr5"             
#>  [82] "MB(+ACA)(R)"       "AL-DL1(R)"         "NO2(L)"           
#>  [85] "SAD(-AMMC)"        "AL-DL3(R)"         "AL-DM5(R)"        
#>  [88] "AL-VL2p(R)"        "AL-VM4(R)"         "NO1(L)"           
#>  [91] "AL-DA4m(L)"        "AL-DP1m(L)"        "PB(R2)"           
#>  [94] "AL-VL2a(R)"        "GF(R)"             "ME(R)"            
#>  [97] "PB(L9)"            "RUB(L)"            "lACA(R)"          
#> [100] "SCL(L)"            "AL-DM3(L)"         "AL-VA1v(R)"       
#> [103] "AL-VA6(R)"         "AL-VP1l(R)"        "NO(R)"            
#> [106] "GOR(R)"            "VES(L)"            "EBr3am"           
#> [109] "AL-VP4(R)"         "a'3(R)"            "PB(R6)"           
#> [112] "AOTU(R)"           "EBr3d"             "AL-DC1(R)"        
#> [115] "SAD"               "b'2(R)"            "SIP(L)"           
#> [118] "AL-DL5(L)"         "AL-VC2(R)"         "AL-DC2(L)"        
#> [121] "FBl6"              "AL-DA3(R)"         "SCL(R)"           
#> [124] "PB(R8)"            "a'2(R)"            "AL-VA6(L)"        
#> [127] "WED(R)"            "POC"               "b'1(R)"           
#> [130] "EPA(R)"            "AL-VM7d(L)"        "FBl3"             
#> [133] "CRE(-RUB)(L)"      "OL(R)"             "GOR(L)"           
#> [136] "PB(L1)"            "PB(L8)"            "SNP(R)"           
#> [139] "AL-VP5(R)"         "hemibrain"         "aL(R)"            
#> [142] "AL-DM1(L)"         "AL-DM2(L)"         "LO(R)"            
#> [145] "AL-DM3(R)"         "AL-VM1(R)"         "AL-VA3(R)"        
#> [148] "VLNP(R)"           "AMMC"              "PB(R7)"           
#> [151] "PB(R1)"            "PB(R4)"            "AL-DA4m(R)"       
#> [154] "AL-DL4(L)"         "a'1(R)"            "PB(L3)"           
#> [157] "AOT(R)"            "AL-DA3(L)"         "ICL(L)"           
#> [160] "AB(R)"             "AL-VA7l(R)"        "AL-DM6(R)"        
#> [163] "AL-VA2(R)"         "AL-DM5(L)"         "AL-DA4l(R)"       
#> [166] "AL-DC2(R)"         "AL-DC3(R)"         "SNP(L)"           
#> [169] "SLP(R)"            "AL-VM3(R)"         "PRW"              
#> [172] "GC"                "EBr3pw"            "PB(L7)"           
#> [175] "MB(L)"             "AL-VA1d(R)"        "NO3(R)"           
#> [178] "NO1(R)"            "PB"                "AL-DC1(L)"        
#> [181] "AL-VC4(R)"         "SMP(R)"            "g3(R)"            
#> [184] "NO3(L)"            "EPA(L)"            "AL(R)"            
#> [187] "AL-VC1(R)"         "CA(L)"             "LAL(-GA)(R)"      
#> [190] "EBr6"              "CA(R)"             "AL-DA1(R)"        
#> [193] "CRE(R)"            "AL-D(L)"           "AL-VA5(R)"        
#> [196] "LAL(L)"            "dACA(R)"           "PB(L2)"           
#> [199] "AL-DM2(R)"         "AL-VL1(R)"         "CX"               
#> [202] "SMP(L)"            "BU(L)"             "gL(L)"            
#> [205] "NO"                "INP"               "BU(R)"            
#> [208] "FB-column3"        "PB(R3)"            "aL(L)"            
#> [211] "AL-VM7v(L)"        "FBl4"              "AL-VP2(R)"        
#> [214] "vACA(R)"           "EB"                "ROB(R)"           
#> [217] "a'L(L)"            "LOP(R)"            "AL-V(R)"          
#> [220] "PVLP(R)"           "FBl9"              "IPS(R)"           
#> [223] "AL-DA2(R)"         "mALT(R)"           "ICL(R)"           
#> [226] "AL-VM2(R)"         "g2(R)"             "AL-DM4(L)"        
#> [229] "AME(R)"            "NO(L)"            
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                    
#> [1,] "GET"  "/api/cached/roiconnectivity"           
#> [2,] "POST" "/api/raw/cypher/cypher"                
#> [3,] "POST" "/api/npexplorer/neuronmeta"            
#> [4,] "GET"  "/api/roimeshes/mesh/:dataset/:roi"     
#> [5,] "POST" "/api/raw/cypher/transaction/:id/cypher"
#> [6,] "GET"  "/token"                                
#>      name                                                                                      
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROIConnectivity-fm"  
#> [2,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                  
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMeta-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/roimeshes.masterAPI.getMesh-fm"          
#> [5,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                  
#> [6,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                  

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }