Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-VA7l(R)"        "AL-DL4(R)"         "AME(R)"           
#>   [4] "FBl1"              "GNG"               "SLP(R)"           
#>   [7] "SCL(L)"            "AL-DM4(L)"         "PB"               
#>  [10] "CX"                "SCL(R)"            "AL-VM7v(L)"       
#>  [13] "AL-VA6(R)"         "AL-VC3m(R)"        "AL-VC4(R)"        
#>  [16] "VLNP(R)"           "PB(L5)"            "dACA(R)"          
#>  [19] "SPS(L)"            "g5(R)"             "NO1(L)"           
#>  [22] "AVLP(R)"           "AL-DL5(R)"         "GA(R)"            
#>  [25] "AL-VC2(R)"         "g2(R)"             "AL-DM1(L)"        
#>  [28] "PB(R9)"            "AL-DL3(R)"         "AL-DL5(L)"        
#>  [31] "AL-VA5(R)"         "AL-VA7m(R)"        "a'3(R)"           
#>  [34] "AL-DC4(L)"         "AL-VM5d(R)"        "bL(L)"            
#>  [37] "AL-VA2(R)"         "LO(R)"             "EBr3am"           
#>  [40] "AL-VP1l(R)"        "AOT(R)"            "BU(L)"            
#>  [43] "AL-DP1m(L)"        "FBl6"              "LH(R)"            
#>  [46] "AL-DC2(R)"         "CRE(-ROB,-RUB)(R)" "SPS(R)"           
#>  [49] "AL-DL4(L)"         "NO(R)"             "LX(R)"            
#>  [52] "AL-DA3(R)"         "PB(L2)"            "VES(L)"           
#>  [55] "a'2(R)"            "EBr2r4"            "b1(R)"            
#>  [58] "AL(R)"             "AL-VA1d(R)"        "AL-V(R)"          
#>  [61] "GC"                "FB-column3"        "LAL(R)"           
#>  [64] "AL-VM1(R)"         "ATL(L)"            "lACA(R)"          
#>  [67] "PB(L9)"            "PB(R4)"            "mALT(L)"          
#>  [70] "ME(R)"             "b'2(R)"            "AL-DA4m(R)"       
#>  [73] "PENP"              "a3(R)"             "AL-VM2(R)"        
#>  [76] "INP"               "CRE(R)"            "ROB(R)"           
#>  [79] "AL-DL2v(R)"        "AL-DM4(R)"         "AL-DP1l(R)"       
#>  [82] "PB(R7)"            "aL(L)"             "AL-DM5(R)"        
#>  [85] "RUB(R)"            "NO2(L)"            "AOTU(R)"          
#>  [88] "AL-DM5(L)"         "NO3(L)"            "SAD(-AMMC)"       
#>  [91] "AL-DC1(R)"         "BU(R)"             "AL-VL1(R)"        
#>  [94] "AMMC"              "AL-DM3(L)"         "bL(R)"            
#>  [97] "FLA(R)"            "ICL(R)"            "b'1(R)"           
#> [100] "hemibrain"         "EBr3pw"            "PB(R2)"           
#> [103] "RUB(L)"            "AL-DM3(R)"         "AL-VC1(R)"        
#> [106] "gL(L)"             "AL-DM2(L)"         "PB(L4)"           
#> [109] "WED(R)"            "NO2(R)"            "CA(R)"            
#> [112] "AL-VP5(R)"         "PB(R6)"            "SAD"              
#> [115] "PED(R)"            "OL(R)"             "SMP(L)"           
#> [118] "b'L(R)"            "EPA(R)"            "ICL(L)"           
#> [121] "LX(L)"             "AL-DM1(R)"         "NO3(R)"           
#> [124] "AL-VC5(R)"         "EBr1"              "AL-DA3(L)"        
#> [127] "GOR(R)"            "LAL(-GA)(R)"       "SIP(R)"           
#> [130] "CRE(L)"            "IB112"             "AL-VA1v(R)"       
#> [133] "AL-VC3l(R)"        "AL-VM7d(R)"        "a1(R)"            
#> [136] "VMNP"              "g1(R)"             "EBr6"             
#> [139] "NO1(R)"            "CRE(-RUB)(L)"      "AL-DA2(R)"        
#> [142] "AL-VA4(R)"         "AL-VP2(R)"         "PRW"              
#> [145] "FB"                "b'L(L)"            "LAL(L)"           
#> [148] "PB(R5)"            "AB(L)"             "LOP(R)"           
#> [151] "PB(L3)"            "AL-DA4m(L)"        "g4(R)"            
#> [154] "FBl3"              "SNP(R)"            "VES(R)"           
#> [157] "PVLP(R)"           "AL-VL2p(R)"        "AL-VP1m(R)"       
#> [160] "CA(L)"             "SIP(L)"            "NO"               
#> [163] "PB(R3)"            "AL-DM2(R)"         "AL-VM3(R)"        
#> [166] "GOR(L)"            "FBl7"              "AL-DC3(R)"        
#> [169] "AL-VA6(L)"         "AL(L)"             "FBl5"             
#> [172] "FBl9"              "EPA(L)"            "AL-VM4(R)"        
#> [175] "AL-VP3(R)"         "g3(R)"             "PB(R8)"           
#> [178] "a'L(R)"            "a'L(L)"            "POC"              
#> [181] "mALT(R)"           "b2(R)"             "a'1(R)"           
#> [184] "AB(R)"             "NO(L)"             "AL-DC2(L)"        
#> [187] "PB(L6)"            "SNP(L)"            "AL-VL2a(R)"       
#> [190] "PLP(R)"            "AL-DC1(L)"         "FBl4"             
#> [193] "AL-DP1m(R)"        "AL-VA3(R)"         "EB"               
#> [196] "PB(R1)"            "IB"                "EBr5"             
#> [199] "gL(R)"             "FBl8"              "AL-DM6(L)"        
#> [202] "ATL(R)"            "AL-DA1(R)"         "MB(+ACA)(R)"      
#> [205] "AL-DA2(L)"         "AL-DL1(R)"         "AL-VM5v(R)"       
#> [208] "AL-VM7v(R)"        "PB(L1)"            "AL-D(R)"          
#> [211] "MB(L)"             "a2(R)"             "PB(L8)"           
#> [214] "EBr3d"             "AL-VM7d(L)"        "PB(L7)"           
#> [217] "IPS(R)"            "AL-VP4(R)"         "GF(R)"            
#> [220] "SMP(R)"            "vACA(R)"           "AL-VP1d(R)"       
#> [223] "AL-DC4(R)"         "AL-DM6(R)"         "aL(R)"            
#> [226] "MB(R)"             "AL-DL2d(R)"        "FBl2"             
#> [229] "CAN(R)"            "AL-D(L)"           "AL-DA4l(R)"       
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "g3(R)"             "GNG"               "VLNP(R)"          
#>   [4] "PB(L2)"            "SPS(R)"            "b'L(R)"           
#>   [7] "PB(R9)"            "g5(R)"             "PB(R5)"           
#>  [10] "SAD"               "VES(R)"            "OL(R)"            
#>  [13] "LOP(R)"            "VMNP"              "ICL(R)"           
#>  [16] "SMP(L)"            "PLP(R)"            "GOR(R)"           
#>  [19] "PB(L5)"            "PB(R7)"            "EBr6"             
#>  [22] "PVLP(R)"           "FBl1"              "AMMC"             
#>  [25] "LH(R)"             "IB"                "AL(L)"            
#>  [28] "ICL(L)"            "MB(+ACA)(R)"       "FBl2"             
#>  [31] "PB(L4)"            "MB(L)"             "PB(L8)"           
#>  [34] "LO(R)"             "CA(L)"             "FBl7"             
#>  [37] "FBl8"              "LAL(L)"            "a2(R)"            
#>  [40] "NO1(L)"            "SCL(R)"            "FBl5"             
#>  [43] "PED(R)"            "g4(R)"             "ROB(R)"           
#>  [46] "aL(L)"             "CRE(-ROB,-RUB)(R)" "EBr2r4"           
#>  [49] "vACA(R)"           "NO1(R)"            "NO3(L)"           
#>  [52] "EBr3pw"            "POC"               "AOT(R)"           
#>  [55] "PB(R4)"            "SLP(R)"            "AB(L)"            
#>  [58] "PENP"              "g2(R)"             "a'L(R)"           
#>  [61] "gL(L)"             "IPS(R)"            "g1(R)"            
#>  [64] "BU(R)"             "ATL(R)"            "NO2(R)"           
#>  [67] "GF(R)"             "dACA(R)"           "AL(R)"            
#>  [70] "SNP(R)"            "PB(L3)"            "GA(R)"            
#>  [73] "SIP(L)"            "CRE(R)"            "EBr3d"            
#>  [76] "b'1(R)"            "mALT(L)"           "NO2(L)"           
#>  [79] "PRW"               "a'1(R)"            "bL(L)"            
#>  [82] "AME(R)"            "LAL(R)"            "CX"               
#>  [85] "PB(R6)"            "NO3(R)"            "EBr5"             
#>  [88] "LX(L)"             "MB(R)"             "GOR(L)"           
#>  [91] "SPS(L)"            "SIP(R)"            "NO(R)"            
#>  [94] "PB(R3)"            "EBr1"              "PB(L7)"           
#>  [97] "CAN(R)"            "EPA(L)"            "a'L(L)"           
#> [100] "a1(R)"             "FB"                "PB(L1)"           
#> [103] "NO"                "PB(L9)"            "bL(R)"            
#> [106] "mALT(R)"           "aL(R)"             "EPA(R)"           
#> [109] "VES(L)"            "ATL(L)"            "lACA(R)"          
#> [112] "BU(L)"             "PB(R8)"            "PB(R1)"           
#> [115] "gL(R)"             "INP"               "WED(R)"           
#> [118] "PB"                "FBl3"              "FBl6"             
#> [121] "FBl9"              "CRE(L)"            "SAD(-AMMC)"       
#> [124] "FBl4"              "b'2(R)"            "LAL(-GA)(R)"      
#> [127] "GC"                "ME(R)"             "EB"               
#> [130] "CA(R)"             "PB(R2)"            "hemibrain"        
#> [133] "AB(R)"             "SMP(R)"            "SCL(L)"           
#> [136] "FLA(R)"            "AOTU(R)"           "NO(L)"            
#> [139] "RUB(R)"            "EBr3am"            "b'L(L)"           
#> [142] "b2(R)"             "LX(R)"             "a'2(R)"           
#> [145] "RUB(L)"            "a3(R)"             "PB(L6)"           
#> [148] "SNP(L)"            "AVLP(R)"           "b1(R)"            
#> [151] "a'3(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DC4(L)"         "AL-VM7d(L)"        "CRE(L)"           
#>   [4] "FBl3"              "AL-DA2(R)"         "AL-DM1(R)"        
#>   [7] "AL-DP1m(R)"        "BU(L)"             "ICL(L)"           
#>  [10] "AL-VA3(R)"         "POC"               "ATL(L)"           
#>  [13] "AOTU(R)"           "dACA(R)"           "AL-VM5v(R)"       
#>  [16] "AB(L)"             "a'3(R)"            "CX"               
#>  [19] "EBr1"              "NO1(L)"            "b2(R)"            
#>  [22] "gL(R)"             "a'1(R)"            "LH(R)"            
#>  [25] "AL(R)"             "AL-DM6(R)"         "g3(R)"            
#>  [28] "AL-VM4(R)"         "MB(L)"             "AL-VC2(R)"        
#>  [31] "PB(R5)"            "mALT(R)"           "IPS(R)"           
#>  [34] "IB"                "LAL(L)"            "VES(L)"           
#>  [37] "SMP(R)"            "AL-DM3(L)"         "AL-VM7v(R)"       
#>  [40] "FBl4"              "AL-DL2d(R)"        "CRE(-ROB,-RUB)(R)"
#>  [43] "PB(L6)"            "SPS(R)"            "CA(L)"            
#>  [46] "SAD"               "a3(R)"             "AL-VM1(R)"        
#>  [49] "AL-VP3(R)"         "PB(L8)"            "PED(R)"           
#>  [52] "SPS(L)"            "AL-VA2(R)"         "PB(R6)"           
#>  [55] "bL(L)"             "AL-DM6(L)"         "AL-VC5(R)"        
#>  [58] "SNP(R)"            "FBl7"              "PB(R7)"           
#>  [61] "NO3(R)"            "LX(R)"             "a'L(R)"           
#>  [64] "AL-DM4(L)"         "PB(R2)"            "RUB(R)"           
#>  [67] "hemibrain"         "AME(R)"            "AL-VA6(L)"        
#>  [70] "b'1(R)"            "FB-column3"        "CAN(R)"           
#>  [73] "AL-VC3l(R)"        "PENP"              "AVLP(R)"          
#>  [76] "RUB(L)"            "bL(R)"             "AL-DC3(R)"        
#>  [79] "AL-VL2p(R)"        "ME(R)"             "BU(R)"            
#>  [82] "AL(L)"             "AL-DM5(L)"         "a'L(L)"           
#>  [85] "AL-DC2(R)"         "AL-VP1m(R)"        "PB(L2)"           
#>  [88] "WED(R)"            "ICL(R)"            "b'L(L)"           
#>  [91] "AL-VM5d(R)"        "EBr3am"            "ROB(R)"           
#>  [94] "PB(R4)"            "AL-VA4(R)"         "VMNP"             
#>  [97] "g5(R)"             "FBl1"              "ATL(R)"           
#> [100] "EPA(L)"            "AL-VP1d(R)"        "vACA(R)"          
#> [103] "PB(L5)"            "PRW"               "AL-V(R)"          
#> [106] "mALT(L)"           "PB(L4)"            "AL-DL5(L)"        
#> [109] "AL-VA6(R)"         "AOT(R)"            "AL-DA2(L)"        
#> [112] "AL-DL5(R)"         "PB(R1)"            "FLA(R)"           
#> [115] "LAL(R)"            "SCL(R)"            "AB(R)"            
#> [118] "AL-DM5(R)"         "EPA(R)"            "CA(R)"            
#> [121] "AL-DC1(R)"         "a'2(R)"            "AL-D(L)"          
#> [124] "PB(L9)"            "FBl9"              "INP"              
#> [127] "MB(R)"             "AL-VA7m(R)"        "AL-VL2a(R)"       
#> [130] "CRE(R)"            "NO2(R)"            "AL-VM3(R)"        
#> [133] "NO3(L)"            "AL-VA1d(R)"        "AL-DC1(L)"        
#> [136] "AL-DM2(R)"         "AL-VP2(R)"         "AL-VP1l(R)"       
#> [139] "NO(L)"             "a1(R)"             "FB"               
#> [142] "aL(R)"             "LAL(-GA)(R)"       "AL-VP4(R)"        
#> [145] "NO(R)"             "GA(R)"             "g2(R)"            
#> [148] "AL-DA3(L)"         "LX(L)"             "OL(R)"            
#> [151] "b'2(R)"            "AL-DC2(L)"         "PB(R3)"           
#> [154] "AL-VC1(R)"         "PB(L1)"            "AL-DM1(L)"        
#> [157] "SLP(R)"            "AL-DC4(R)"         "g4(R)"            
#> [160] "NO"                "PB(L7)"            "AL-DM2(L)"        
#> [163] "SAD(-AMMC)"        "LOP(R)"            "EBr3pw"           
#> [166] "g1(R)"             "AL-VM2(R)"         "lACA(R)"          
#> [169] "PB(R8)"            "EBr3d"             "AL-DP1m(L)"       
#> [172] "MB(+ACA)(R)"       "GOR(R)"            "AL-VA1v(R)"       
#> [175] "PB(L3)"            "AL-VM7d(R)"        "AL-VA5(R)"        
#> [178] "PVLP(R)"           "gL(L)"             "b'L(R)"           
#> [181] "SIP(L)"            "AL-DA4m(L)"        "FBl5"             
#> [184] "AL-DL4(R)"         "GOR(L)"            "SIP(R)"           
#> [187] "SNP(L)"            "FBl6"              "AL-DM4(R)"        
#> [190] "AL-VC3m(R)"        "SMP(L)"            "VES(R)"           
#> [193] "AL-DM3(R)"         "AL-VL1(R)"         "NO1(R)"           
#> [196] "AL-DA1(R)"         "AL-DL1(R)"         "AL-D(R)"          
#> [199] "AL-VC4(R)"         "FBl2"              "EB"               
#> [202] "VLNP(R)"           "PB"                "AL-VM7v(L)"       
#> [205] "aL(L)"             "AL-DL3(R)"         "AL-VP5(R)"        
#> [208] "PLP(R)"            "AL-DA4l(R)"        "NO2(L)"           
#> [211] "a2(R)"             "AL-DA4m(R)"        "EBr6"             
#> [214] "EBr5"              "GNG"               "GF(R)"            
#> [217] "EBr2r4"            "AMMC"              "GC"               
#> [220] "PB(R9)"            "FBl8"              "AL-DL2v(R)"       
#> [223] "AL-DL4(L)"         "LO(R)"             "CRE(-RUB)(L)"     
#> [226] "AL-DA3(R)"         "b1(R)"             "AL-VA7l(R)"       
#> [229] "SCL(L)"            "AL-DP1l(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "FBl4"              "AL-VM7d(R)"        "NO3(R)"           
#>   [4] "FB"                "SNP(R)"            "AL(L)"            
#>   [7] "ATL(R)"            "SIP(L)"            "AL-DM3(R)"        
#>  [10] "AL-DM5(R)"         "NO"                "AL-DL3(R)"        
#>  [13] "CA(L)"             "AOTU(R)"           "EPA(R)"           
#>  [16] "AL-VA2(R)"         "AL-VC3l(R)"        "NO2(L)"           
#>  [19] "g1(R)"             "dACA(R)"           "AL-DA2(L)"        
#>  [22] "mALT(L)"           "AL-DC1(L)"         "bL(L)"            
#>  [25] "AL-DA4m(L)"        "PB(L3)"            "FBl7"             
#>  [28] "AL(R)"             "NO3(L)"            "PLP(R)"           
#>  [31] "LO(R)"             "NO2(R)"            "OL(R)"            
#>  [34] "AL-DP1m(R)"        "LAL(L)"            "FLA(R)"           
#>  [37] "AL-DM6(L)"         "AL-VL1(R)"         "AL-VP3(R)"        
#>  [40] "PB(R7)"            "PED(R)"            "g2(R)"            
#>  [43] "EBr3am"            "GOR(L)"            "GC"               
#>  [46] "FB-column3"        "AL-DA1(R)"         "AL-VA3(R)"        
#>  [49] "NO1(L)"            "AL-V(R)"           "WED(R)"           
#>  [52] "AL-DL5(R)"         "AL-VA7l(R)"        "AL-VC5(R)"        
#>  [55] "EBr3d"             "AL-DC4(L)"         "AL-VM7v(R)"       
#>  [58] "AL-DP1m(L)"        "AL-VA6(L)"         "PB(R2)"           
#>  [61] "AL-DC1(R)"         "SNP(L)"            "SAD"              
#>  [64] "VES(R)"            "PB(L4)"            "SCL(L)"           
#>  [67] "CRE(-ROB,-RUB)(R)" "AL-DL4(L)"         "CA(R)"            
#>  [70] "MB(R)"             "AL-DL4(R)"         "AL-D(R)"          
#>  [73] "AL-VA4(R)"         "AL-VA5(R)"         "PB(L1)"           
#>  [76] "NO(R)"             "AL-DL1(R)"         "AL-DL2v(R)"       
#>  [79] "AL-VP4(R)"         "AL-DL5(L)"         "aL(L)"            
#>  [82] "a'L(L)"            "CX"                "b'2(R)"           
#>  [85] "gL(L)"             "FBl6"              "AL-DC2(R)"        
#>  [88] "LOP(R)"            "b2(R)"             "EBr1"             
#>  [91] "AL-VM5d(R)"        "POC"               "AVLP(R)"          
#>  [94] "AL-DC3(R)"         "AL-VP1d(R)"        "CRE(R)"           
#>  [97] "EB"                "EPA(L)"            "PB(L5)"           
#> [100] "LAL(-GA)(R)"       "AL-DM2(L)"         "AME(R)"           
#> [103] "NO(L)"             "SPS(L)"            "AL-D(L)"          
#> [106] "AL-VA1v(R)"        "AL-VM5v(R)"        "EBr5"             
#> [109] "AL-DM5(L)"         "FBl1"              "FBl5"             
#> [112] "AL-DL2d(R)"        "lACA(R)"           "AB(R)"            
#> [115] "a3(R)"             "PENP"              "LX(R)"            
#> [118] "AL-VC1(R)"         "AL-VL2p(R)"        "a'2(R)"           
#> [121] "BU(R)"             "ICL(R)"            "AL-VP1l(R)"       
#> [124] "PB(R5)"            "FBl2"              "VES(L)"           
#> [127] "VLNP(R)"           "AOT(R)"            "BU(L)"            
#> [130] "PB(L8)"            "SIP(R)"            "MB(+ACA)(R)"      
#> [133] "AL-VL2a(R)"        "AL-VM3(R)"         "RUB(R)"           
#> [136] "SMP(L)"            "PRW"               "a'L(R)"           
#> [139] "b'L(L)"            "CAN(R)"            "AL-VM1(R)"        
#> [142] "SPS(R)"            "PB"                "AL-DM4(L)"        
#> [145] "FBl3"              "PB(L6)"            "a1(R)"            
#> [148] "AL-DM1(L)"         "FBl8"              "AMMC"             
#> [151] "PB(R8)"            "EBr2r4"            "RUB(L)"           
#> [154] "AL-DM6(R)"         "AL-VA7m(R)"        "GNG"              
#> [157] "LH(R)"             "AL-DA4m(R)"        "AL-VA6(R)"        
#> [160] "AL-DA3(L)"         "ROB(R)"            "bL(R)"            
#> [163] "AL-DC4(R)"         "AL-VC3m(R)"        "GA(R)"            
#> [166] "b'L(R)"            "NO1(R)"            "SLP(R)"           
#> [169] "AL-DA2(R)"         "AL-DM1(R)"         "AL-DM2(R)"        
#> [172] "AL-DC2(L)"         "a'1(R)"            "LX(L)"            
#> [175] "a2(R)"             "AL-VC2(R)"         "AL-VM4(R)"        
#> [178] "SMP(R)"            "PB(R1)"            "SAD(-AMMC)"       
#> [181] "FBl9"              "AL-VP5(R)"         "gL(R)"            
#> [184] "PB(L9)"            "GOR(R)"            "AL-DA4l(R)"       
#> [187] "AL-VM7d(L)"        "ATL(L)"            "MB(L)"            
#> [190] "EBr6"              "IPS(R)"            "mALT(R)"          
#> [193] "ME(R)"             "g4(R)"             "CRE(-RUB)(L)"     
#> [196] "PB(L7)"            "PVLP(R)"           "AL-VP2(R)"        
#> [199] "hemibrain"         "aL(R)"             "AL-VM7v(L)"       
#> [202] "AL-VA1d(R)"        "b'1(R)"            "EBr3pw"           
#> [205] "PB(R3)"            "PB(R4)"            "LAL(R)"           
#> [208] "IB"                "PB(R6)"            "SCL(R)"           
#> [211] "AL-DM3(L)"         "CRE(L)"            "INP"              
#> [214] "AL-DM4(R)"         "vACA(R)"           "PB(R9)"           
#> [217] "AL-DA3(R)"         "AL-DP1l(R)"        "AL-VP1m(R)"       
#> [220] "g3(R)"             "AB(L)"             "a'3(R)"           
#> [223] "VMNP"              "ICL(L)"            "GF(R)"            
#> [226] "g5(R)"             "b1(R)"             "AL-VC4(R)"        
#> [229] "AL-VM2(R)"         "PB(L2)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method     path                                
#> [1,] "POST"     "/api/v:ver/npexplorer/rankedtable" 
#> [2,] "PROPFIND" "/api/*"                            
#> [3,] "GET"      "/*"                                
#> [4,] "GET"      "/api/v:ver/cached/roiconnectivity" 
#> [5,] "PATCH"    "/api"                              
#> [6,] "POST"     "/api/v:ver/npexplorer/roisinneuron"
#>      name                                                                
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [2,] "github.com/labstack/echo.glob..func1"                              
#> [3,] "github.com/labstack/echo.common.static.func1"                      
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [5,] "github.com/labstack/echo.glob..func1"                              
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }