Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "PB(L9)"            "AME(R)"            "AL-DM2(R)"        
#>   [4] "RUB(R)"            "GA(R)"             "MB(L)"            
#>   [7] "AMMC"              "AL-DL5(L)"         "b1(R)"            
#>  [10] "AL-VC2(R)"         "PB(R5)"            "AL-DC3(R)"        
#>  [13] "SCL(R)"            "FBl7"              "b'L(L)"           
#>  [16] "AL-DA1(R)"         "PENP"              "aL(L)"            
#>  [19] "SMP(R)"            "SNP(R)"            "EBr5"             
#>  [22] "EPA(L)"            "AL-DC2(R)"         "AL-VM7d(L)"       
#>  [25] "VES(R)"            "INP"               "AL-DA3(L)"        
#>  [28] "a'2(R)"            "FBl1"              "CAN(R)"           
#>  [31] "a'3(R)"            "CRE(-RUB)(L)"      "LX(L)"            
#>  [34] "AL-VA4(R)"         "PRW"               "SIP(L)"           
#>  [37] "AL-DM6(L)"         "PB(R3)"            "EB"               
#>  [40] "PVLP(R)"           "AL-DL3(R)"         "AL-VA5(R)"        
#>  [43] "AL-VC1(R)"         "AL-VL2a(R)"        "AL-VP4(R)"        
#>  [46] "SMP(L)"            "a3(R)"             "AL-DC4(R)"        
#>  [49] "AL-D(R)"           "AL-VP2(R)"         "AL-VA7l(R)"       
#>  [52] "AL-VL2p(R)"        "PB"                "AL-DA4m(R)"       
#>  [55] "LX(R)"             "AVLP(R)"           "AL-VP3(R)"        
#>  [58] "PB(L3)"            "a'L(R)"            "AL-DC2(L)"        
#>  [61] "AL-VM7d(R)"        "PLP(R)"            "AL-DA2(L)"        
#>  [64] "AL-DL1(R)"         "AL-DL5(R)"         "PB(L8)"           
#>  [67] "RUB(L)"            "POC"               "PB(R4)"           
#>  [70] "a1(R)"             "AOTU(R)"           "ATL(R)"           
#>  [73] "AL-DM4(R)"         "AL-VM5d(R)"        "AL-DM5(L)"        
#>  [76] "AL-VC4(R)"         "AL-VM7v(R)"        "AL-DM1(R)"        
#>  [79] "AL-DP1l(R)"        "VLNP(R)"           "PB(R6)"           
#>  [82] "SIP(R)"            "IB112"             "IPS(R)"           
#>  [85] "EBr3am"            "AL-DC1(L)"         "NO(R)"            
#>  [88] "AL-VC5(R)"         "AL-VL1(R)"         "VMNP"             
#>  [91] "BU(R)"             "OL(R)"             "LOP(R)"           
#>  [94] "hemibrain"         "AL-DA2(R)"         "PB(L6)"           
#>  [97] "SAD(-AMMC)"        "AOT(R)"            "GNG"              
#> [100] "AL-DL2v(R)"        "GF(R)"             "EBr3pw"           
#> [103] "MB(+ACA)(R)"       "NO2(R)"            "a2(R)"            
#> [106] "FBl3"              "AL-VM3(R)"         "NO3(R)"           
#> [109] "NO3(L)"            "LAL(R)"            "CRE(-ROB,-RUB)(R)"
#> [112] "PED(R)"            "EBr2r4"            "EPA(R)"           
#> [115] "FBl8"              "CA(L)"             "mALT(L)"          
#> [118] "WED(R)"            "a'1(R)"            "VES(L)"           
#> [121] "AL-DM5(R)"         "g1(R)"             "AL(L)"            
#> [124] "AL-DM2(L)"         "PB(L5)"            "PB(L7)"           
#> [127] "NO1(L)"            "EBr3d"             "gL(R)"            
#> [130] "GC"                "AL-DM3(L)"         "g3(R)"            
#> [133] "AB(R)"             "AL-DM3(R)"         "b'2(R)"           
#> [136] "SLP(R)"            "AL-DL2d(R)"        "AL-VP1l(R)"       
#> [139] "SCL(L)"            "PB(R7)"            "AL-VA6(L)"        
#> [142] "CRE(L)"            "AL-VM2(R)"         "AL-DA4m(L)"       
#> [145] "AL-D(L)"           "FBl5"              "a'L(L)"           
#> [148] "AL-VM4(R)"         "FBl6"              "dACA(R)"          
#> [151] "PB(L4)"            "AL-VP1m(R)"        "g4(R)"            
#> [154] "GOR(R)"            "b2(R)"             "FBl2"             
#> [157] "SAD"               "b'L(R)"            "PB(R2)"           
#> [160] "ATL(L)"            "ME(R)"             "PB(L1)"           
#> [163] "PB(R8)"            "g5(R)"             "AL-VA1d(R)"       
#> [166] "PB(L2)"            "mALT(R)"           "AL-DM4(L)"        
#> [169] "AL(R)"             "AL-DA4l(R)"        "SPS(L)"           
#> [172] "AL-VP5(R)"         "AL-V(R)"           "AL-VM7v(L)"       
#> [175] "ICL(R)"            "AL-VA3(R)"         "AL-VC3l(R)"       
#> [178] "LO(R)"             "FB-column3"        "CA(R)"            
#> [181] "AL-VA7m(R)"        "FBl9"              "AL-DM6(R)"        
#> [184] "AL-VA2(R)"         "AB(L)"             "SPS(R)"           
#> [187] "lACA(R)"           "AL-DP1m(L)"        "ROB(R)"           
#> [190] "CX"                "EBr6"              "AL-DL4(L)"        
#> [193] "AL-DP1m(R)"        "IB"                "NO"               
#> [196] "FBl4"              "LH(R)"             "AL-VA6(R)"        
#> [199] "LAL(L)"            "g2(R)"             "FLA(R)"           
#> [202] "AL-DC4(L)"         "AL-VM1(R)"         "SNP(L)"           
#> [205] "NO(L)"             "MB(R)"             "AL-VA1v(R)"       
#> [208] "GOR(L)"            "gL(L)"             "PB(R1)"           
#> [211] "ICL(L)"            "NO1(R)"            "bL(R)"            
#> [214] "AL-DA3(R)"         "AL-VP1d(R)"        "FB"               
#> [217] "EBr1"              "b'1(R)"            "AL-VM5v(R)"       
#> [220] "vACA(R)"           "aL(R)"             "AL-DM1(L)"        
#> [223] "bL(L)"             "AL-DC1(R)"         "LAL(-GA)(R)"      
#> [226] "BU(L)"             "PB(R9)"            "AL-VC3m(R)"       
#> [229] "CRE(R)"            "AL-DL4(R)"         "NO2(L)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "g3(R)"             "ME(R)"             "LO(R)"            
#>   [4] "RUB(R)"            "LAL(L)"            "SPS(R)"           
#>   [7] "lACA(R)"           "b'L(R)"            "a'L(R)"           
#>  [10] "FBl2"              "AB(R)"             "AOTU(R)"          
#>  [13] "PED(R)"            "LAL(-GA)(R)"       "g5(R)"            
#>  [16] "FBl7"              "FBl8"              "PB(R2)"           
#>  [19] "AMMC"              "PB(R6)"            "MB(L)"            
#>  [22] "SNP(R)"            "g4(R)"             "FBl5"             
#>  [25] "CAN(R)"            "a3(R)"             "EBr6"             
#>  [28] "BU(L)"             "PB"                "ATL(L)"           
#>  [31] "CA(L)"             "b'2(R)"            "dACA(R)"          
#>  [34] "PB(R8)"            "BU(R)"             "AL(L)"            
#>  [37] "CA(R)"             "b1(R)"             "NO2(R)"           
#>  [40] "PB(R5)"            "a'3(R)"            "PB(L5)"           
#>  [43] "mALT(L)"           "CRE(R)"            "IPS(R)"           
#>  [46] "SLP(R)"            "g1(R)"             "VMNP"             
#>  [49] "ROB(R)"            "FBl1"              "INP"              
#>  [52] "mALT(R)"           "VES(R)"            "b2(R)"            
#>  [55] "FLA(R)"            "NO(R)"             "AL(R)"            
#>  [58] "EBr3am"            "SPS(L)"            "PB(L6)"           
#>  [61] "SAD(-AMMC)"        "vACA(R)"           "PB(L3)"           
#>  [64] "EPA(R)"            "MB(+ACA)(R)"       "CX"               
#>  [67] "EBr3d"             "FBl4"              "EPA(L)"           
#>  [70] "GOR(R)"            "SCL(L)"            "PB(L1)"           
#>  [73] "PLP(R)"            "gL(L)"             "LX(L)"            
#>  [76] "a'L(L)"            "GC"                "NO3(L)"           
#>  [79] "EBr3pw"            "SNP(L)"            "bL(R)"            
#>  [82] "OL(R)"             "FB"                "EBr2r4"           
#>  [85] "PB(L9)"            "PVLP(R)"           "GNG"              
#>  [88] "ICL(R)"            "RUB(L)"            "NO(L)"            
#>  [91] "GA(R)"             "aL(R)"             "SCL(R)"           
#>  [94] "NO"                "PB(L7)"            "PB(R4)"           
#>  [97] "a2(R)"             "hemibrain"         "SIP(L)"           
#> [100] "b'L(L)"            "NO2(L)"            "NO3(R)"           
#> [103] "PB(L8)"            "gL(R)"             "LOP(R)"           
#> [106] "PENP"              "PB(R7)"            "EBr5"             
#> [109] "AVLP(R)"           "LAL(R)"            "AOT(R)"           
#> [112] "SMP(L)"            "EBr1"              "VES(L)"           
#> [115] "MB(R)"             "a1(R)"             "PRW"              
#> [118] "PB(R9)"            "a'1(R)"            "FBl9"             
#> [121] "LH(R)"             "VLNP(R)"           "CRE(-ROB,-RUB)(R)"
#> [124] "GOR(L)"            "WED(R)"            "a'2(R)"           
#> [127] "PB(R3)"            "POC"               "SIP(R)"           
#> [130] "AME(R)"            "b'1(R)"            "NO1(L)"           
#> [133] "g2(R)"             "EB"                "PB(R1)"           
#> [136] "CRE(L)"            "FBl6"              "aL(L)"            
#> [139] "IB"                "NO1(R)"            "LX(R)"            
#> [142] "ATL(R)"            "PB(L4)"            "SMP(R)"           
#> [145] "PB(L2)"            "GF(R)"             "FBl3"             
#> [148] "AB(L)"             "SAD"               "bL(L)"            
#> [151] "ICL(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "CX"                "gL(L)"             "AL-VA6(L)"        
#>   [4] "RUB(L)"            "FLA(R)"            "SAD(-AMMC)"       
#>   [7] "FBl6"              "FBl8"              "AL-DP1m(R)"       
#>  [10] "AL-VL2p(R)"        "EBr3pw"            "FB-column3"       
#>  [13] "AL-VC1(R)"         "AL-VP5(R)"         "AL-VP3(R)"        
#>  [16] "AL-VP1m(R)"        "AL-DA2(L)"         "AL-DM6(L)"        
#>  [19] "VES(L)"            "EBr3am"            "ATL(L)"           
#>  [22] "SMP(L)"            "AL-VC5(R)"         "EBr3d"            
#>  [25] "a3(R)"             "AL-VL1(R)"         "MB(L)"            
#>  [28] "AL-DC2(L)"         "PB(R3)"            "AL-VC4(R)"        
#>  [31] "AL-VP2(R)"         "VLNP(R)"           "EBr5"             
#>  [34] "AVLP(R)"           "AL-DA2(R)"         "AL-DM3(R)"        
#>  [37] "AL-VA1d(R)"        "AL-VA2(R)"         "mALT(R)"          
#>  [40] "PB(L7)"            "a1(R)"             "GC"               
#>  [43] "AL(R)"             "AL-DC2(R)"         "AL-VM7d(L)"       
#>  [46] "NO"                "AL-DL1(R)"         "AL-DM5(R)"        
#>  [49] "AL-VM2(R)"         "b'1(R)"            "AL(L)"            
#>  [52] "AL-VM3(R)"         "b'2(R)"            "LAL(L)"           
#>  [55] "FB"                "AB(L)"             "AOTU(R)"          
#>  [58] "PB(R7)"            "AL-VA7m(R)"        "NO2(L)"           
#>  [61] "dACA(R)"           "PVLP(R)"           "AL-D(R)"          
#>  [64] "PB(L5)"            "AL-DM1(L)"         "PB(R9)"           
#>  [67] "AL-DL4(L)"         "AL-DM5(L)"         "AL-DM1(R)"        
#>  [70] "AL-VA3(R)"         "RUB(R)"            "CRE(-ROB,-RUB)(R)"
#>  [73] "AL-DC4(L)"         "b'L(L)"            "AL-VM5d(R)"       
#>  [76] "AL-V(R)"           "POC"               "AL-DL3(R)"        
#>  [79] "SCL(R)"            "SIP(R)"            "PB(R1)"           
#>  [82] "EPA(L)"            "bL(R)"             "AL-DC1(R)"        
#>  [85] "LX(R)"             "AL-DM2(R)"         "AL-DA3(L)"        
#>  [88] "CRE(R)"            "AL-DM3(L)"         "FBl3"             
#>  [91] "FBl7"              "AL-VM7v(R)"        "AL-VP4(R)"        
#>  [94] "LH(R)"             "AL-VC3l(R)"        "SIP(L)"           
#>  [97] "SPS(R)"            "LO(R)"             "CAN(R)"           
#> [100] "AL-VA1v(R)"        "LOP(R)"            "EBr2r4"           
#> [103] "NO3(L)"            "MB(+ACA)(R)"       "OL(R)"            
#> [106] "aL(R)"             "NO3(R)"            "PB(R8)"           
#> [109] "GOR(R)"            "lACA(R)"           "PB(R5)"           
#> [112] "CA(L)"             "SPS(L)"            "AL-DC1(L)"        
#> [115] "FBl5"              "AL-VM5v(R)"        "PB(L8)"           
#> [118] "NO1(R)"            "AL-DM4(L)"         "CRE(L)"           
#> [121] "FBl9"              "ATL(R)"            "a'L(R)"           
#> [124] "ROB(R)"            "MB(R)"             "AL-DA1(R)"        
#> [127] "AL-DA3(R)"         "PB(R2)"            "IPS(R)"           
#> [130] "AME(R)"            "AL-DA4l(R)"        "AL-DL2d(R)"       
#> [133] "b2(R)"             "EB"                "AL-VA7l(R)"       
#> [136] "PENP"              "GA(R)"             "SAD"              
#> [139] "PB(R4)"            "SCL(L)"            "BU(R)"            
#> [142] "FBl4"              "hemibrain"         "PB(L6)"           
#> [145] "AL-VM7d(R)"        "NO1(L)"            "AL-DA4m(L)"       
#> [148] "AL-DC4(R)"         "AL-VP1l(R)"        "PB(R6)"           
#> [151] "GOR(L)"            "WED(R)"            "PB(L1)"           
#> [154] "PB(L9)"            "VES(R)"            "vACA(R)"          
#> [157] "PRW"               "PLP(R)"            "g4(R)"            
#> [160] "GNG"               "NO(L)"             "LAL(R)"           
#> [163] "AL-DM4(R)"         "AL-DM6(R)"         "IB"               
#> [166] "GF(R)"             "bL(L)"             "INP"              
#> [169] "AL-D(L)"           "b1(R)"             "PED(R)"           
#> [172] "EBr6"              "a'1(R)"            "SNP(R)"           
#> [175] "gL(R)"             "BU(L)"             "a'2(R)"           
#> [178] "EPA(R)"            "AL-DC3(R)"         "AL-VP1d(R)"       
#> [181] "EBr1"              "AL-DP1m(L)"        "NO2(R)"           
#> [184] "AL-DA4m(R)"        "LAL(-GA)(R)"       "VMNP"             
#> [187] "AL-VA5(R)"         "CRE(-RUB)(L)"      "ICL(L)"           
#> [190] "g2(R)"             "aL(L)"             "AL-VA6(R)"        
#> [193] "AOT(R)"            "AMMC"              "a2(R)"            
#> [196] "FBl1"              "SLP(R)"            "AL-DL5(R)"        
#> [199] "AL-VL2a(R)"        "AL-DM2(L)"         "AL-VM7v(L)"       
#> [202] "AL-DL4(R)"         "g1(R)"             "b'L(R)"           
#> [205] "g5(R)"             "ICL(R)"            "PB(L3)"           
#> [208] "AL-DL5(L)"         "NO(R)"             "AL-DP1l(R)"       
#> [211] "SNP(L)"            "AB(R)"             "PB(L4)"           
#> [214] "AL-DL2v(R)"        "ME(R)"             "AL-VC2(R)"        
#> [217] "CA(R)"             "PB(L2)"            "PB"               
#> [220] "LX(L)"             "AL-VA4(R)"         "AL-VC3m(R)"       
#> [223] "AL-VM4(R)"         "a'3(R)"            "g3(R)"            
#> [226] "SMP(R)"            "FBl2"              "a'L(L)"           
#> [229] "AL-VM1(R)"         "mALT(L)"          
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "FB-column3"        "AL-DL1(R)"         "AL-VL1(R)"        
#>   [4] "AL-VM4(R)"         "SNP(L)"            "VES(R)"           
#>   [7] "CA(R)"             "PB(R4)"            "dACA(R)"          
#>  [10] "gL(R)"             "b'L(L)"            "LAL(R)"           
#>  [13] "AL-DM1(R)"         "AL-V(R)"           "GOR(L)"           
#>  [16] "LAL(-GA)(R)"       "FB"                "bL(L)"            
#>  [19] "AL-VM5v(R)"        "a2(R)"             "SLP(R)"           
#>  [22] "AL-DA4m(R)"        "AL-DL2d(R)"        "VLNP(R)"          
#>  [25] "AL-DA3(L)"         "AL-VP1l(R)"        "IB"               
#>  [28] "b'L(R)"            "FBl9"              "AL-DC1(R)"        
#>  [31] "AL-DM5(R)"         "AL-VM7v(R)"        "mALT(R)"          
#>  [34] "ATL(R)"            "EPA(L)"            "AL-DA2(R)"        
#>  [37] "AL-DC4(R)"         "AL-VC5(R)"         "AL-DL5(L)"        
#>  [40] "AL-VA2(R)"         "FBl7"              "AL-VA6(L)"        
#>  [43] "AL-DL4(R)"         "AL-DM6(R)"         "AB(L)"            
#>  [46] "AL-DP1m(L)"        "SAD(-AMMC)"        "GF(R)"            
#>  [49] "a1(R)"             "WED(R)"            "g4(R)"            
#>  [52] "AL-DL2v(R)"        "g2(R)"             "AL-D(L)"          
#>  [55] "AL-DM1(L)"         "AL-VC3m(R)"        "NO3(R)"           
#>  [58] "AL(R)"             "AL-VA1d(R)"        "AL-VA5(R)"        
#>  [61] "AL-VM2(R)"         "CRE(-ROB,-RUB)(R)" "FBl1"             
#>  [64] "INP"               "AL-VA3(R)"         "AL-DL4(L)"        
#>  [67] "CRE(-RUB)(L)"      "AME(R)"            "AL-VP5(R)"        
#>  [70] "AL-VP1m(R)"        "b2(R)"             "VMNP"             
#>  [73] "AL-VM7d(L)"        "NO3(L)"            "PB(L1)"           
#>  [76] "NO(R)"             "AMMC"              "gL(L)"            
#>  [79] "AL-VM7d(R)"        "BU(L)"             "AL-DC2(L)"        
#>  [82] "bL(R)"             "AL-DL5(R)"         "CRE(R)"           
#>  [85] "a'1(R)"            "LH(R)"             "AL-VC4(R)"        
#>  [88] "PB(L5)"            "PB(R6)"            "AL-DM2(L)"        
#>  [91] "SMP(R)"            "RUB(R)"            "lACA(R)"          
#>  [94] "AL-VP2(R)"         "LAL(L)"            "SPS(R)"           
#>  [97] "b'1(R)"            "g3(R)"             "g5(R)"            
#> [100] "AL-DA4l(R)"        "AL-DM4(R)"         "AL-VP4(R)"        
#> [103] "PB(R9)"            "PB(L7)"            "GOR(R)"           
#> [106] "OL(R)"             "AL-DC3(R)"         "AL-VP3(R)"        
#> [109] "SNP(R)"            "AL-DA1(R)"         "AL-DA3(R)"        
#> [112] "AL-DM3(L)"         "AL-DM6(L)"         "AL-VM1(R)"        
#> [115] "AL-VM5d(R)"        "EBr5"              "AL-VL2p(R)"       
#> [118] "CAN(R)"            "AL-VA4(R)"         "PB"               
#> [121] "ROB(R)"            "MB(R)"             "AL-DA2(L)"        
#> [124] "SMP(L)"            "AL-VM7v(L)"        "AL-DC2(R)"        
#> [127] "AL-VM3(R)"         "b'2(R)"            "PB(L9)"           
#> [130] "PRW"               "FBl3"              "a'L(L)"           
#> [133] "AL-D(R)"           "EBr3am"            "ME(R)"            
#> [136] "a'2(R)"            "FBl6"              "AVLP(R)"          
#> [139] "AL-VC3l(R)"        "AL-VP1d(R)"        "vACA(R)"          
#> [142] "NO2(L)"            "PB(L3)"            "NO1(R)"           
#> [145] "a'L(R)"            "b1(R)"             "AL-DP1m(R)"       
#> [148] "PB(L2)"            "PB(R7)"            "FBl2"             
#> [151] "PB(R3)"            "PB(L4)"            "CX"               
#> [154] "FLA(R)"            "NO"                "GNG"              
#> [157] "AL-VA7l(R)"        "LX(R)"             "AL-DM5(L)"        
#> [160] "AL-DP1l(R)"        "AL-VL2a(R)"        "AL-DC1(L)"        
#> [163] "EBr1"              "GC"                "AL-DC4(L)"        
#> [166] "RUB(L)"            "AL-VC1(R)"         "AOT(R)"           
#> [169] "SCL(L)"            "SAD"               "FBl8"             
#> [172] "AL-DM4(L)"         "MB(+ACA)(R)"       "PLP(R)"           
#> [175] "EPA(R)"            "mALT(L)"           "aL(R)"            
#> [178] "EB"                "CA(L)"             "hemibrain"        
#> [181] "SIP(L)"            "AOTU(R)"           "AB(R)"            
#> [184] "AL-DM3(R)"         "PB(R1)"            "PB(R2)"           
#> [187] "LX(L)"             "a3(R)"             "AL-VC2(R)"        
#> [190] "SCL(R)"            "PB(L6)"            "IPS(R)"           
#> [193] "aL(L)"             "AL-VA7m(R)"        "SIP(R)"           
#> [196] "PB(L8)"            "CRE(L)"            "PVLP(R)"          
#> [199] "FBl5"              "NO1(L)"            "BU(R)"            
#> [202] "PB(R5)"            "ATL(L)"            "PED(R)"           
#> [205] "EBr3d"             "AL(L)"             "POC"              
#> [208] "a'3(R)"            "PB(R8)"            "EBr2r4"           
#> [211] "FBl4"              "ICL(R)"            "VES(L)"           
#> [214] "AL-VA1v(R)"        "SPS(L)"            "MB(L)"            
#> [217] "g1(R)"             "AL-DM2(R)"         "PENP"             
#> [220] "AL-VA6(R)"         "NO(L)"             "AL-DA4m(L)"       
#> [223] "NO2(R)"            "EBr6"              "LO(R)"            
#> [226] "LOP(R)"            "GA(R)"             "ICL(L)"           
#> [229] "AL-DL3(R)"         "EBr3pw"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                     
#> [1,] "GET"  "/api/npexplorer/roisinneuron"           
#> [2,] "POST" "/api/npexplorer/autapses"               
#> [3,] "GET"  "/api/custom/custom"                     
#> [4,] "POST" "/api/v:ver/npexplorer/simpleconnections"
#> [5,] "GET"  "/api/v:ver/npexplorer/autapses"         
#> [6,] "GET"  "/api/npexplorer/celltype/:dataset/:type"
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"    
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"             
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCellType-fm"    
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }