Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "EBr2r4"            "lACA(R)"           "WED(R)"           
#>   [4] "dACA(R)"           "FBl3"              "EBr3am"           
#>   [7] "NO1(R)"            "ME(R)"             "g4(R)"            
#>  [10] "EBr3pw"            "PB(L7)"            "MB(+ACA)(R)"      
#>  [13] "PB(R7)"            "mALT(R)"           "EBr1"             
#>  [16] "VES(R)"            "RUB(R)"            "AB(L)"            
#>  [19] "PB(R8)"            "BU(R)"             "AL(L)"            
#>  [22] "LO(R)"             "NO2(R)"            "EPA(L)"           
#>  [25] "SPS(R)"            "GC"                "LX(L)"            
#>  [28] "RUB(L)"            "NO(L)"             "PB(L2)"           
#>  [31] "NO3(L)"            "a2(R)"             "SAD(-AMMC)"       
#>  [34] "PB(R4)"            "a'3(R)"            "g3(R)"            
#>  [37] "NO(R)"             "FBl5"              "FBl9"             
#>  [40] "NO3(R)"            "CRE(R)"            "ATL(R)"           
#>  [43] "AB(R)"             "SLP(R)"            "LAL(L)"           
#>  [46] "INP"               "GNG"               "NO1(L)"           
#>  [49] "FLA(R)"            "PLP(R)"            "PB(R9)"           
#>  [52] "PB(R1)"            "ICL(L)"            "PB(R2)"           
#>  [55] "CA(L)"             "aL(R)"             "LX(R)"            
#>  [58] "a'L(R)"            "AVLP(R)"           "AME(R)"           
#>  [61] "b'1(R)"            "CX"                "PB(L3)"           
#>  [64] "GA(R)"             "SCL(R)"            "PB(L1)"           
#>  [67] "FBl1"              "SMP(R)"            "b'2(R)"           
#>  [70] "SMP(L)"            "GOR(L)"            "LAL(-GA)(R)"      
#>  [73] "a'2(R)"            "EB"                "CRE(L)"           
#>  [76] "b2(R)"             "PB(L8)"            "AL(R)"            
#>  [79] "CRE(-ROB,-RUB)(R)" "AOTU(R)"           "BU(L)"            
#>  [82] "EPA(R)"            "SIP(R)"            "gL(R)"            
#>  [85] "aL(L)"             "PB(R5)"            "hemibrain"        
#>  [88] "SAD"               "g1(R)"             "IPS(R)"           
#>  [91] "a3(R)"             "POC"               "SCL(L)"           
#>  [94] "PB(L5)"            "AMMC"              "NO2(L)"           
#>  [97] "bL(R)"             "IB"                "PENP"             
#> [100] "b'L(R)"            "FBl8"              "a'L(L)"           
#> [103] "VES(L)"            "AOT(R)"            "PB(R6)"           
#> [106] "SIP(L)"            "PB(L9)"            "FBl4"             
#> [109] "PB(R3)"            "OL(R)"             "vACA(R)"          
#> [112] "mALT(L)"           "PRW"               "g2(R)"            
#> [115] "g5(R)"             "FBl6"              "CAN(R)"           
#> [118] "SNP(L)"            "ICL(R)"            "b1(R)"            
#> [121] "LAL(R)"            "gL(L)"             "LH(R)"            
#> [124] "MB(R)"             "PED(R)"            "b'L(L)"           
#> [127] "EBr6"              "ROB(R)"            "FBl7"             
#> [130] "GOR(R)"            "PB(L4)"            "VLNP(R)"          
#> [133] "SNP(R)"            "NO"                "CA(R)"            
#> [136] "MB(L)"             "SPS(L)"            "VMNP"             
#> [139] "FBl2"              "LOP(R)"            "PB(L6)"           
#> [142] "PB"                "EBr3d"             "PVLP(R)"          
#> [145] "ATL(L)"            "a1(R)"             "GF(R)"            
#> [148] "FB"                "EBr5"              "a'1(R)"           
#> [151] "bL(L)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method  path                            
#> [1,] "GET"   "/api/npexplorer/neuronmetavals"
#> [2,] "GET"   "/api/cached/dailytype"         
#> [3,] "GET"   "/api/*"                        
#> [4,] "TRACE" "/api/*"                        
#> [5,] "GET"   "/api/dbmeta/datasets"          
#> [6,] "POST"  "/api/npexplorer/neuronmetavals"
#>      name                                                                                          
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMetaVals-fm"
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getDailyType-fm"            
#> [3,] "github.com/labstack/echo.glob..func1"                                                        
#> [4,] "github.com/labstack/echo.glob..func1"                                                        
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatasets-fm"              
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMetaVals-fm"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }