Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DC2(L)"         "hemibrain"         "a'L(R)"           
#>   [4] "AL-DM1(R)"         "AL-D(R)"           "AL-VA3(R)"        
#>   [7] "RUB(R)"            "PB(R6)"            "VMNP"             
#>  [10] "FBl5"              "NO2(R)"            "MB(R)"            
#>  [13] "SMP(L)"            "MB(+ACA)(R)"       "PB(R9)"           
#>  [16] "PB(R3)"            "AL-VM5d(R)"        "mALT(R)"          
#>  [19] "AL-DC3(R)"         "AL-DA2(L)"         "ICL(L)"           
#>  [22] "g1(R)"             "EBr1"              "PB(R2)"           
#>  [25] "EPA(L)"            "PB(L4)"            "AL-VP4(R)"        
#>  [28] "PB(R5)"            "LAL(-GA)(R)"       "NO1(R)"           
#>  [31] "VES(L)"            "NO3(R)"            "PB(R8)"           
#>  [34] "AL-DM4(R)"         "NO(R)"             "a'1(R)"           
#>  [37] "LAL(L)"            "g5(R)"             "AL-DL3(R)"        
#>  [40] "AL-DL2v(R)"        "LX(R)"             "AL-DP1m(L)"       
#>  [43] "AL-DL5(R)"         "AL-VA7m(R)"        "AL-DL4(L)"        
#>  [46] "AL-DL1(R)"         "IB112"             "VLNP(R)"          
#>  [49] "IB"                "EBr3d"             "SNP(L)"           
#>  [52] "AL-VC4(R)"         "CX"                "CRE(R)"           
#>  [55] "GOR(R)"            "AMMC"              "PB(L6)"           
#>  [58] "bL(L)"             "b'L(L)"            "AB(L)"            
#>  [61] "EBr5"              "BU(L)"             "AL-DA4l(R)"       
#>  [64] "AL-DC4(R)"         "AL-VA5(R)"         "AL-VP1m(R)"       
#>  [67] "g4(R)"             "SAD(-AMMC)"        "AL-VA7l(R)"       
#>  [70] "LOP(R)"            "PB(L5)"            "CRE(-ROB,-RUB)(R)"
#>  [73] "PB(L1)"            "AL-DA2(R)"         "AOT(R)"           
#>  [76] "AL-DA3(L)"         "a3(R)"             "SNP(R)"           
#>  [79] "ATL(R)"            "PRW"               "SIP(L)"           
#>  [82] "FBl2"              "AL-VA1d(R)"        "AL-DM2(L)"        
#>  [85] "SIP(R)"            "AL-DA3(R)"         "GC"               
#>  [88] "EBr3pw"            "CA(R)"             "OL(R)"            
#>  [91] "AL-VL1(R)"         "AL-VP2(R)"         "b'L(R)"           
#>  [94] "AL-VA2(R)"         "AL-VM1(R)"         "AL-VA6(L)"        
#>  [97] "PB(R4)"            "AL-DC2(R)"         "b2(R)"            
#> [100] "PB"                "AL-DM3(L)"         "AL-DL2d(R)"       
#> [103] "ME(R)"             "BU(R)"             "PB(R7)"           
#> [106] "mALT(L)"           "AL-DM1(L)"         "VES(R)"           
#> [109] "AL-DM4(L)"         "AL-VP1l(R)"        "AL-VM7d(L)"       
#> [112] "PB(L9)"            "CRE(L)"            "ICL(R)"           
#> [115] "RUB(L)"            "GNG"               "WED(R)"           
#> [118] "FLA(R)"            "FBl1"              "NO1(L)"           
#> [121] "EPA(R)"            "EBr3am"            "FB"               
#> [124] "AL-VM4(R)"         "AL-DM3(R)"         "AL-VC3m(R)"       
#> [127] "GA(R)"             "LX(L)"             "AL-DP1m(R)"       
#> [130] "AL-V(R)"           "ATL(L)"            "PLP(R)"           
#> [133] "NO(L)"             "g3(R)"             "AL-VC5(R)"        
#> [136] "AL-VC3l(R)"        "AL-VP1d(R)"        "POC"              
#> [139] "FBl4"              "AME(R)"            "PB(R1)"           
#> [142] "CAN(R)"            "AL-VP3(R)"         "a'3(R)"           
#> [145] "PB(L8)"            "ROB(R)"            "AL-DA4m(L)"       
#> [148] "g2(R)"             "EBr2r4"            "AL-VC2(R)"        
#> [151] "AL-VM2(R)"         "b'2(R)"            "LH(R)"            
#> [154] "SAD"               "a'L(L)"            "b1(R)"            
#> [157] "AL-VL2a(R)"        "a1(R)"             "PENP"             
#> [160] "FBl8"              "AL-VM7v(L)"        "LAL(R)"           
#> [163] "AL-DM5(R)"         "lACA(R)"           "AVLP(R)"          
#> [166] "AL-DC1(L)"         "AL-VM7v(R)"        "NO2(L)"           
#> [169] "AOTU(R)"           "dACA(R)"           "IPS(R)"           
#> [172] "INP"               "SCL(R)"            "AB(R)"            
#> [175] "FBl7"              "AL-DL4(R)"         "AL-DM5(L)"        
#> [178] "gL(R)"             "EBr6"              "AL-DL5(L)"        
#> [181] "FB-column3"        "AL-VA4(R)"         "CRE(-RUB)(L)"     
#> [184] "LO(R)"             "FBl9"              "AL-DM2(R)"        
#> [187] "AL-VA1v(R)"        "AL-VL2p(R)"        "a'2(R)"           
#> [190] "PB(L7)"            "GF(R)"             "vACA(R)"          
#> [193] "AL-DA1(R)"         "GOR(L)"            "SLP(R)"           
#> [196] "AL(R)"             "b'1(R)"            "AL-DC4(L)"        
#> [199] "SPS(L)"            "AL-D(L)"           "AL-DA4m(R)"       
#> [202] "AL-VM5v(R)"        "SPS(R)"            "AL(L)"            
#> [205] "CA(L)"             "AL-DM6(R)"         "SCL(L)"           
#> [208] "FBl6"              "AL-VA6(R)"         "PED(R)"           
#> [211] "PVLP(R)"           "AL-DP1l(R)"        "AL-VM7d(R)"       
#> [214] "EB"                "bL(R)"             "AL-VP5(R)"        
#> [217] "SMP(R)"            "PB(L3)"            "MB(L)"            
#> [220] "NO"                "FBl3"              "aL(L)"            
#> [223] "a2(R)"             "gL(L)"             "NO3(L)"           
#> [226] "AL-DC1(R)"         "AL-VC1(R)"         "AL-VM3(R)"        
#> [229] "PB(L2)"            "aL(R)"             "AL-DM6(L)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(L5)"            "SPS(L)"            "LX(R)"            
#>   [4] "AVLP(R)"           "LX(L)"             "VLNP(R)"          
#>   [7] "SCL(L)"            "PENP"              "b2(R)"            
#>  [10] "PB(R2)"            "b1(R)"             "IB"               
#>  [13] "LOP(R)"            "g1(R)"             "FB"               
#>  [16] "SCL(R)"            "FBl2"              "GC"               
#>  [19] "aL(R)"             "NO"                "CRE(-ROB,-RUB)(R)"
#>  [22] "AMMC"              "dACA(R)"           "SIP(R)"           
#>  [25] "LH(R)"             "aL(L)"             "RUB(R)"           
#>  [28] "vACA(R)"           "LAL(L)"            "FLA(R)"           
#>  [31] "FBl6"              "b'L(R)"            "BU(R)"            
#>  [34] "AL(L)"             "NO(R)"             "RUB(L)"           
#>  [37] "PB(R5)"            "ME(R)"             "SNP(R)"           
#>  [40] "MB(+ACA)(R)"       "ATL(R)"            "IPS(R)"           
#>  [43] "a3(R)"             "NO1(L)"            "AL(R)"            
#>  [46] "OL(R)"             "NO2(L)"            "GA(R)"            
#>  [49] "PED(R)"            "LAL(-GA)(R)"       "PB(L8)"           
#>  [52] "PB(L6)"            "a'1(R)"            "b'2(R)"           
#>  [55] "gL(L)"             "INP"               "PB(R6)"           
#>  [58] "SAD(-AMMC)"        "FBl5"              "mALT(L)"          
#>  [61] "EBr1"              "NO1(R)"            "AB(R)"            
#>  [64] "PB(R4)"            "SMP(R)"            "CRE(R)"           
#>  [67] "BU(L)"             "PB(R3)"            "a'L(L)"           
#>  [70] "SIP(L)"            "PB(L4)"            "NO3(R)"           
#>  [73] "g2(R)"             "NO3(L)"            "PVLP(R)"          
#>  [76] "PB(L2)"            "PB(R7)"            "gL(R)"            
#>  [79] "CAN(R)"            "b'L(L)"            "VES(L)"           
#>  [82] "SLP(R)"            "PB"                "CA(R)"            
#>  [85] "ICL(R)"            "bL(R)"             "EBr6"             
#>  [88] "PB(L7)"            "AB(L)"             "ATL(L)"           
#>  [91] "SMP(L)"            "lACA(R)"           "GF(R)"            
#>  [94] "CRE(L)"            "LO(R)"             "VES(R)"           
#>  [97] "FBl4"              "POC"               "CA(L)"            
#> [100] "a2(R)"             "WED(R)"            "g4(R)"            
#> [103] "EPA(R)"            "FBl7"              "LAL(R)"           
#> [106] "SAD"               "PRW"               "EBr2r4"           
#> [109] "FBl3"              "g3(R)"             "PB(L3)"           
#> [112] "ROB(R)"            "ICL(L)"            "EPA(L)"           
#> [115] "MB(R)"             "hemibrain"         "AME(R)"           
#> [118] "NO(L)"             "AOTU(R)"           "mALT(R)"          
#> [121] "PB(R8)"            "FBl9"              "NO2(R)"           
#> [124] "SPS(R)"            "GOR(L)"            "EBr5"             
#> [127] "g5(R)"             "SNP(L)"            "b'1(R)"           
#> [130] "a1(R)"             "PB(L1)"            "EBr3d"            
#> [133] "EB"                "a'L(R)"            "PB(R9)"           
#> [136] "GOR(R)"            "AOT(R)"            "a'3(R)"           
#> [139] "PLP(R)"            "VMNP"              "a'2(R)"           
#> [142] "bL(L)"             "EBr3pw"            "GNG"              
#> [145] "EBr3am"            "CX"                "MB(L)"            
#> [148] "PB(R1)"            "PB(L9)"            "FBl1"             
#> [151] "FBl8"             
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "GC"                "CRE(-RUB)(L)"      "ATL(R)"           
#>   [4] "AL-VP5(R)"         "SPS(R)"            "SNP(L)"           
#>   [7] "AL-VA1d(R)"        "PB(R4)"            "BU(R)"            
#>  [10] "AL-VP4(R)"         "AL-VP1l(R)"        "PB(L2)"           
#>  [13] "PB"                "AL-DC4(R)"         "AL-DM6(R)"        
#>  [16] "AL-VM2(R)"         "a1(R)"             "AL-DA2(L)"        
#>  [19] "NO2(R)"            "AL-DC1(R)"         "PB(R5)"           
#>  [22] "AL-VC3l(R)"        "NO1(L)"            "EBr3d"            
#>  [25] "SNP(R)"            "AL-DM2(R)"         "PB(R6)"           
#>  [28] "CA(R)"             "RUB(R)"            "b'1(R)"           
#>  [31] "AL-DL3(R)"         "lACA(R)"           "CRE(R)"           
#>  [34] "FB-column3"        "VES(L)"            "CRE(-ROB,-RUB)(R)"
#>  [37] "PB(R7)"            "NO3(R)"            "AL-DM1(L)"        
#>  [40] "AOT(R)"            "AL-VC5(R)"         "VLNP(R)"          
#>  [43] "dACA(R)"           "PB(R8)"            "EBr2r4"           
#>  [46] "PB(L9)"            "PB(L4)"            "AL-DA3(R)"        
#>  [49] "SCL(L)"            "b'2(R)"            "BU(L)"            
#>  [52] "AL-DM3(L)"         "a'1(R)"            "aL(L)"            
#>  [55] "AL-DL2v(R)"        "AL-VC4(R)"         "AL-VL2p(R)"       
#>  [58] "EBr1"              "AL-VM7v(R)"        "AL-VA4(R)"        
#>  [61] "AL-VA5(R)"         "a'3(R)"            "NO2(L)"           
#>  [64] "g3(R)"             "PED(R)"            "a'2(R)"           
#>  [67] "INP"               "AL(R)"             "AL-VA1v(R)"       
#>  [70] "POC"               "PB(L1)"            "PB(L8)"           
#>  [73] "GOR(R)"            "LAL(R)"            "ME(R)"            
#>  [76] "AL-DP1m(R)"        "ICL(R)"            "SLP(R)"           
#>  [79] "AL-VM1(R)"         "AB(L)"             "AL-DM6(L)"        
#>  [82] "AL-DL1(R)"         "AL-VC1(R)"         "WED(R)"           
#>  [85] "FBl1"              "AMMC"              "AL-DM4(L)"        
#>  [88] "GNG"               "PB(L3)"            "hemibrain"        
#>  [91] "CAN(R)"            "PB(R2)"            "mALT(R)"          
#>  [94] "MB(L)"             "gL(R)"             "EBr6"             
#>  [97] "gL(L)"             "AL-DM4(R)"         "AL-VM7d(R)"       
#> [100] "AOTU(R)"           "AL-DC4(L)"         "AL-DL4(L)"        
#> [103] "AL-DM3(R)"         "AL-D(R)"           "AL-VA3(R)"        
#> [106] "AL-VP2(R)"         "aL(R)"             "AL-D(L)"          
#> [109] "AL-VM7v(L)"        "bL(R)"             "AL-VA6(R)"        
#> [112] "AL-VM4(R)"         "AL-DA2(R)"         "AL-VA7m(R)"       
#> [115] "SIP(L)"            "SCL(R)"            "FLA(R)"           
#> [118] "PB(R3)"            "AL-VL1(R)"         "IB"               
#> [121] "EBr5"              "LX(R)"             "AME(R)"           
#> [124] "AL-DM1(R)"         "FBl3"              "AL-DC3(R)"        
#> [127] "AL-V(R)"           "AL-DL4(R)"         "AL-DL5(R)"        
#> [130] "PB(L5)"            "AL-DL5(L)"         "bL(L)"            
#> [133] "b1(R)"             "SMP(L)"            "g4(R)"            
#> [136] "AL-VM7d(L)"        "RUB(L)"            "LH(R)"            
#> [139] "OL(R)"             "vACA(R)"           "AL-VA2(R)"        
#> [142] "EBr3am"            "b2(R)"             "AL-DP1m(L)"       
#> [145] "PB(L7)"            "LO(R)"             "FB"               
#> [148] "CA(L)"             "FBl7"              "EPA(L)"           
#> [151] "a3(R)"             "AL-DM5(R)"         "b'L(R)"           
#> [154] "AL(L)"             "AL-DC2(L)"         "NO"               
#> [157] "AL-VM5v(R)"        "AL-DL2d(R)"        "AL-DA3(L)"        
#> [160] "AL-DP1l(R)"        "AL-DA4m(R)"        "AL-DC2(R)"        
#> [163] "ATL(L)"            "NO(R)"             "LX(L)"            
#> [166] "b'L(L)"            "NO(L)"             "GA(R)"            
#> [169] "AL-VP1d(R)"        "AL-VA7l(R)"        "g2(R)"            
#> [172] "ROB(R)"            "FBl4"              "a'L(L)"           
#> [175] "AL-VC2(R)"         "g5(R)"             "FBl2"             
#> [178] "FBl8"              "FBl9"              "EBr3pw"           
#> [181] "AB(R)"             "LAL(-GA)(R)"       "NO1(R)"           
#> [184] "VMNP"              "AL-DC1(L)"         "FBl6"             
#> [187] "AL-VM3(R)"         "AL-VM5d(R)"        "MB(R)"            
#> [190] "AL-DA1(R)"         "PRW"               "PB(R1)"           
#> [193] "SAD(-AMMC)"        "MB(+ACA)(R)"       "g1(R)"            
#> [196] "EB"                "SIP(R)"            "NO3(L)"           
#> [199] "AL-DA4m(L)"        "AL-VA6(L)"         "mALT(L)"          
#> [202] "GF(R)"             "PLP(R)"            "AL-VP3(R)"        
#> [205] "LOP(R)"            "EPA(R)"            "AL-DM5(L)"        
#> [208] "ICL(L)"            "VES(R)"            "SAD"              
#> [211] "CRE(L)"            "PVLP(R)"           "SPS(L)"           
#> [214] "AL-DM2(L)"         "AL-VC3m(R)"        "SMP(R)"           
#> [217] "AVLP(R)"           "AL-DA4l(R)"        "CX"               
#> [220] "PENP"              "GOR(L)"            "a'L(R)"           
#> [223] "PB(L6)"            "LAL(L)"            "a2(R)"            
#> [226] "FBl5"              "AL-VL2a(R)"        "PB(R9)"           
#> [229] "IPS(R)"            "AL-VP1m(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-DM6(R)"         "AL-VC1(R)"         "vACA(R)"          
#>   [4] "AL-DM4(L)"         "PB(R1)"            "IPS(R)"           
#>   [7] "PB(R4)"            "NO2(L)"            "PB(R6)"           
#>  [10] "AL-DC1(R)"         "CX"                "PB(R7)"           
#>  [13] "AL-VM7d(L)"        "AL-DL5(R)"         "b'1(R)"           
#>  [16] "BU(L)"             "g4(R)"             "AL-DA2(R)"        
#>  [19] "AL-DL2d(R)"        "AOT(R)"            "PB(R8)"           
#>  [22] "AL-DM1(L)"         "g5(R)"             "POC"              
#>  [25] "SNP(L)"            "VES(L)"            "AL(R)"            
#>  [28] "AL-VM7d(R)"        "b'2(R)"            "g2(R)"            
#>  [31] "AL-VA4(R)"         "AL-VP1l(R)"        "PB(R5)"           
#>  [34] "a2(R)"             "NO1(R)"            "PB(L4)"           
#>  [37] "AL-DM4(R)"         "mALT(L)"           "g1(R)"            
#>  [40] "b2(R)"             "bL(L)"             "FBl4"             
#>  [43] "AL-DA4m(R)"        "AL-VA6(R)"         "PRW"              
#>  [46] "AL-VA5(R)"         "AMMC"              "b'L(R)"           
#>  [49] "AL-DM3(R)"         "AL-VA2(R)"         "a'L(L)"           
#>  [52] "AL-VL2p(R)"        "SPS(R)"            "mALT(R)"          
#>  [55] "WED(R)"            "EBr6"              "AL-DC4(L)"        
#>  [58] "PB(L7)"            "AL-DM2(R)"         "AL-VC3l(R)"       
#>  [61] "SAD(-AMMC)"        "FBl8"              "LH(R)"            
#>  [64] "SCL(R)"            "AL-DL5(L)"         "AL-DM5(L)"        
#>  [67] "FBl2"              "AL-VP2(R)"         "ME(R)"            
#>  [70] "AL-DL4(L)"         "AL-D(R)"           "AL-VM3(R)"        
#>  [73] "PENP"              "gL(L)"             "AL-DA4m(L)"       
#>  [76] "AL-DM2(L)"         "AL-VA3(R)"         "LOP(R)"           
#>  [79] "dACA(R)"           "NO(L)"             "AL-VC5(R)"        
#>  [82] "AL-VM7v(L)"        "GOR(R)"            "ROB(R)"           
#>  [85] "AL-DC4(R)"         "AL-VP4(R)"         "PB(L3)"           
#>  [88] "PED(R)"            "AL(L)"             "MB(+ACA)(R)"      
#>  [91] "CA(R)"             "AL-VP1d(R)"        "RUB(R)"           
#>  [94] "GOR(L)"            "LAL(-GA)(R)"       "EBr3d"            
#>  [97] "AL-DC1(L)"         "EPA(L)"            "EBr3am"           
#> [100] "a1(R)"             "ICL(L)"            "AL-VM7v(R)"       
#> [103] "SCL(L)"            "FB"                "SIP(R)"           
#> [106] "LX(L)"             "AL-DC3(R)"         "AL-VL1(R)"        
#> [109] "AL-VP5(R)"         "SAD"               "EBr1"             
#> [112] "AL-VA1d(R)"        "SPS(L)"            "EPA(R)"           
#> [115] "AL-D(L)"           "OL(R)"             "AL-VL2a(R)"       
#> [118] "NO1(L)"            "NO"                "PB(L9)"           
#> [121] "VES(R)"            "FBl9"              "LAL(R)"           
#> [124] "AL-DA1(R)"         "AL-DA3(R)"         "IB"               
#> [127] "a'2(R)"            "AL-DP1m(R)"        "EB"               
#> [130] "VMNP"              "SNP(R)"            "PB(L6)"           
#> [133] "AME(R)"            "b'L(L)"            "GF(R)"            
#> [136] "FBl7"              "b1(R)"             "AL-DM1(R)"        
#> [139] "VLNP(R)"           "lACA(R)"           "PB(L1)"           
#> [142] "AL-DA2(L)"         "MB(R)"             "AL-V(R)"          
#> [145] "AL-VC3m(R)"        "PB(L5)"            "NO3(R)"           
#> [148] "AL-DM3(L)"         "AL-DM6(L)"         "AL-DP1m(L)"       
#> [151] "AL-DM5(R)"         "SIP(L)"            "AL-VM2(R)"        
#> [154] "PB(L2)"            "CRE(L)"            "FBl1"             
#> [157] "a'3(R)"            "BU(R)"             "LO(R)"            
#> [160] "FBl3"              "ATL(R)"            "AL-VA1v(R)"       
#> [163] "AL-VA7l(R)"        "AL-VM5d(R)"        "FLA(R)"           
#> [166] "a'L(R)"            "NO(R)"             "CRE(-RUB)(L)"     
#> [169] "RUB(L)"            "AL-VM4(R)"         "NO2(R)"           
#> [172] "AL-DL1(R)"         "CRE(-ROB,-RUB)(R)" "PB(R2)"           
#> [175] "AB(R)"             "bL(R)"             "SLP(R)"           
#> [178] "SMP(L)"            "GNG"               "AL-DL4(R)"        
#> [181] "LX(R)"             "AL-DA3(L)"         "PVLP(R)"          
#> [184] "AL-VP1m(R)"        "ATL(L)"            "INP"              
#> [187] "AL-DL2v(R)"        "AVLP(R)"           "AL-DC2(R)"        
#> [190] "SMP(R)"            "MB(L)"             "AL-VC4(R)"        
#> [193] "CA(L)"             "g3(R)"             "NO3(L)"           
#> [196] "a3(R)"             "LAL(L)"            "PB"               
#> [199] "AL-VA6(L)"         "FBl5"              "aL(L)"            
#> [202] "AL-VA7m(R)"        "AL-VM1(R)"         "AL-VC2(R)"        
#> [205] "PB(L8)"            "gL(R)"             "EBr3pw"           
#> [208] "hemibrain"         "PLP(R)"            "ICL(R)"           
#> [211] "aL(R)"             "GA(R)"             "EBr2r4"           
#> [214] "AL-DL3(R)"         "FB-column3"        "AL-DA4l(R)"       
#> [217] "CRE(R)"            "EBr5"              "AB(L)"            
#> [220] "PB(R9)"            "a'1(R)"            "PB(R3)"           
#> [223] "AOTU(R)"           "CAN(R)"            "AL-VP3(R)"        
#> [226] "GC"                "FBl6"              "AL-DC2(L)"        
#> [229] "AL-DP1l(R)"        "AL-VM5v(R)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                             
#> [1,] "POST" "/api/npexplorer/distribution"   
#> [2,] "GET"  "/login"                         
#> [3,] "PUT"  "/api"                           
#> [4,] "POST" "/api/npexplorer/roiconnectivity"
#> [5,] "POST" "/api/v:ver/raw/cypher/cypher"   
#> [6,] "GET"  "/api/v:ver/npexplorer/autapses" 
#>      name                                                                                           
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"   
#> [2,] "github.com/janelia-flyem/echo-secure.loginHandler"                                            
#> [3,] "github.com/labstack/echo.glob..func1"                                                         
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIConnectivity-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }