Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "g4(R)"             "LO(R)"             "INP"              
#>   [4] "NO(L)"             "PB(L2)"            "EBr1"             
#>   [7] "PRW"               "RUB(L)"            "a2(R)"            
#>  [10] "a'L(R)"            "SAD(-AMMC)"        "AL(R)"            
#>  [13] "PB(L1)"            "WED(R)"            "BU(L)"            
#>  [16] "VMNP"              "FBl6"              "AB(L)"            
#>  [19] "SMP(R)"            "SCL(L)"            "mALT(R)"          
#>  [22] "PB(L5)"            "LAL(-GA)(R)"       "NO1(L)"           
#>  [25] "SCL(R)"            "EBr2r4"            "NO"               
#>  [28] "ICL(L)"            "CAN(R)"            "GF(R)"            
#>  [31] "GC"                "LH(R)"             "MB(R)"            
#>  [34] "SIP(L)"            "PLP(R)"            "SIP(R)"           
#>  [37] "PB(R9)"            "RUB(R)"            "PB"               
#>  [40] "PB(L8)"            "SNP(R)"            "NO(R)"            
#>  [43] "b'L(L)"            "aL(R)"             "GOR(L)"           
#>  [46] "PB(R8)"            "AB(R)"             "a'L(L)"           
#>  [49] "OL(R)"             "lACA(R)"           "bL(R)"            
#>  [52] "PB(L4)"            "EBr6"              "SLP(R)"           
#>  [55] "a'1(R)"            "EBr3pw"            "PB(L7)"           
#>  [58] "FBl2"              "b1(R)"             "NO2(R)"           
#>  [61] "LAL(R)"            "a'2(R)"            "a1(R)"            
#>  [64] "b'2(R)"            "b'L(R)"            "EPA(R)"           
#>  [67] "PB(R6)"            "FBl9"              "CA(L)"            
#>  [70] "ME(R)"             "AL(L)"             "NO3(L)"           
#>  [73] "GOR(R)"            "VLNP(R)"           "POC"              
#>  [76] "IB"                "PED(R)"            "AOTU(R)"          
#>  [79] "VES(R)"            "LAL(L)"            "PB(R5)"           
#>  [82] "EBr3am"            "vACA(R)"           "PB(R4)"           
#>  [85] "FB"                "SPS(L)"            "FBl7"             
#>  [88] "AME(R)"            "GNG"               "AMMC"             
#>  [91] "NO2(L)"            "MB(L)"             "BU(R)"            
#>  [94] "SNP(L)"            "CX"                "NO1(R)"           
#>  [97] "g1(R)"             "EBr5"              "FLA(R)"           
#> [100] "PB(L6)"            "ATL(R)"            "AVLP(R)"          
#> [103] "aL(L)"             "PB(L3)"            "CRE(-ROB,-RUB)(R)"
#> [106] "PB(L9)"            "CRE(L)"            "PVLP(R)"          
#> [109] "MB(+ACA)(R)"       "PB(R2)"            "a'3(R)"           
#> [112] "PB(R7)"            "FBl5"              "CA(R)"            
#> [115] "VES(L)"            "g3(R)"             "LX(L)"            
#> [118] "GA(R)"             "bL(L)"             "FBl1"             
#> [121] "FBl3"              "EPA(L)"            "a3(R)"            
#> [124] "LOP(R)"            "gL(R)"             "b2(R)"            
#> [127] "g2(R)"             "LX(R)"             "gL(L)"            
#> [130] "FBl4"              "SAD"               "SPS(R)"           
#> [133] "dACA(R)"           "g5(R)"             "SMP(L)"           
#> [136] "ROB(R)"            "FBl8"              "PENP"             
#> [139] "PB(R1)"            "AOT(R)"            "b'1(R)"           
#> [142] "NO3(R)"            "ICL(R)"            "CRE(R)"           
#> [145] "ATL(L)"            "EBr3d"             "IPS(R)"           
#> [148] "hemibrain"         "PB(R3)"            "mALT(L)"          
#> [151] "EB"               
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                        
#> [1,] "POST" "/api/v:ver/npexplorer/completeness"        
#> [2,] "POST" "/api/skeletons/skeleton/:dataset/:id"      
#> [3,] "GET"  "/api/roimeshes/mesh/:dataset/:roi"         
#> [4,] "POST" "/api/v:ver/raw/cypher/transaction/:id/kill"
#> [5,] "GET"  "/api/cached/roicompleteness"               
#> [6,] "GET"  "/api/npexplorer/roisinneuron"              
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [2,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/roimeshes.masterAPI.getMesh-fm"            
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROICompleteness-fm"    
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }