Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "g5(R)"             "NO2(R)"            "g3(R)"            
#>   [4] "PRW"               "PB(R8)"            "PB"               
#>   [7] "hemibrain"         "SNP(R)"            "b1(R)"            
#>  [10] "SLP(R)"            "PB(L9)"            "SNP(L)"           
#>  [13] "PB(R3)"            "PB(L2)"            "FBl3"             
#>  [16] "ICL(R)"            "GOR(L)"            "GC"               
#>  [19] "EBr6"              "NO3(L)"            "a'L(R)"           
#>  [22] "VES(R)"            "ATL(R)"            "aL(L)"            
#>  [25] "WED(R)"            "NO3(R)"            "EBr5"             
#>  [28] "VMNP"              "GNG"               "LOP(R)"           
#>  [31] "AMMC"              "EPA(R)"            "NO(R)"            
#>  [34] "MB(R)"             "PB(L8)"            "NO(L)"            
#>  [37] "IB"                "CRE(R)"            "FB"               
#>  [40] "a'2(R)"            "PB(L6)"            "CRE(L)"           
#>  [43] "AME(R)"            "NO2(L)"            "GA(R)"            
#>  [46] "dACA(R)"           "NO1(R)"            "NO"               
#>  [49] "mALT(L)"           "LAL(L)"            "lACA(R)"          
#>  [52] "aL(R)"             "gL(L)"             "SAD(-AMMC)"       
#>  [55] "FBl2"              "EPA(L)"            "AB(L)"            
#>  [58] "AL(L)"             "PB(R9)"            "ROB(R)"           
#>  [61] "PB(L7)"            "PENP"              "PB(R1)"           
#>  [64] "LX(L)"             "CA(L)"             "GF(R)"            
#>  [67] "AVLP(R)"           "vACA(R)"           "IPS(R)"           
#>  [70] "CAN(R)"            "AOT(R)"            "ATL(L)"           
#>  [73] "b'2(R)"            "b'L(R)"            "PLP(R)"           
#>  [76] "SIP(R)"            "CRE(-ROB,-RUB)(R)" "PB(R6)"           
#>  [79] "SAD"               "EBr3d"             "PB(L4)"           
#>  [82] "FLA(R)"            "FBl4"              "CA(R)"            
#>  [85] "RUB(R)"            "g2(R)"             "GOR(R)"           
#>  [88] "POC"               "PB(R5)"            "FBl7"             
#>  [91] "VLNP(R)"           "bL(L)"             "LO(R)"            
#>  [94] "MB(+ACA)(R)"       "FBl1"              "AL(R)"            
#>  [97] "b2(R)"             "EBr2r4"            "RUB(L)"           
#> [100] "a3(R)"             "EBr1"              "BU(L)"            
#> [103] "a'1(R)"            "SMP(L)"            "PB(R2)"           
#> [106] "LAL(R)"            "PB(R7)"            "SMP(R)"           
#> [109] "a1(R)"             "ME(R)"             "b'L(L)"           
#> [112] "SCL(L)"            "EBr3am"            "EBr3pw"           
#> [115] "FBl9"              "PB(R4)"            "SIP(L)"           
#> [118] "SCL(R)"            "SPS(L)"            "BU(R)"            
#> [121] "NO1(L)"            "ICL(L)"            "OL(R)"            
#> [124] "LAL(-GA)(R)"       "MB(L)"             "PB(L5)"           
#> [127] "mALT(R)"           "g4(R)"             "gL(R)"            
#> [130] "INP"               "SPS(R)"            "PVLP(R)"          
#> [133] "LH(R)"             "bL(R)"             "PB(L1)"           
#> [136] "EB"                "CX"                "FBl8"             
#> [139] "AB(R)"             "VES(L)"            "b'1(R)"           
#> [142] "PED(R)"            "g1(R)"             "LX(R)"            
#> [145] "FBl5"              "a'3(R)"            "a'L(L)"           
#> [148] "PB(L3)"            "AOTU(R)"           "a2(R)"            
#> [151] "FBl6"             
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-VC2(R)"         "PB(L2)"            "PB(L8)"           
#>   [4] "FBl8"              "AL-DP1m(R)"        "AL-VM7d(R)"       
#>   [7] "mALT(L)"           "PB(R1)"            "INP"              
#>  [10] "AL-VM4(R)"         "CRE(-ROB,-RUB)(R)" "ATL(L)"           
#>  [13] "SCL(R)"            "VMNP"              "AL-VC3l(R)"       
#>  [16] "aL(L)"             "AL-DA4m(R)"        "AL-VP1l(R)"       
#>  [19] "PLP(R)"            "AL-DL4(L)"         "ATL(R)"           
#>  [22] "g2(R)"             "EBr3d"             "AL-DL2v(R)"       
#>  [25] "AL-DM3(R)"         "AL-DM4(R)"         "AL-D(L)"          
#>  [28] "EBr3pw"            "AL-DL1(R)"         "GOR(R)"           
#>  [31] "AL-VA3(R)"         "AL-V(R)"           "AL-VP1d(R)"       
#>  [34] "SMP(R)"            "AL-DC1(L)"         "LAL(R)"           
#>  [37] "AL-DL4(R)"         "AL-VP2(R)"         "LX(R)"            
#>  [40] "PB(L6)"            "SAD(-AMMC)"        "AL-VA4(R)"        
#>  [43] "SNP(R)"            "a'L(L)"            "AL-DL2d(R)"       
#>  [46] "AL-VC5(R)"         "CA(R)"             "PB(L4)"           
#>  [49] "AL-DC4(R)"         "AL-DA2(L)"         "SLP(R)"           
#>  [52] "NO1(R)"            "AL-VA2(R)"         "AL-DL5(L)"        
#>  [55] "AL-VC4(R)"         "AL-VM7v(R)"        "b'L(R)"           
#>  [58] "ICL(R)"            "AL-DP1l(R)"        "AL-DM1(R)"        
#>  [61] "VLNP(R)"           "PB(L9)"            "PB(R2)"           
#>  [64] "IB"                "PENP"              "FBl1"             
#>  [67] "PB(R4)"            "b1(R)"             "AL-DC2(R)"        
#>  [70] "AL-VA1d(R)"        "AL-VM2(R)"         "a'3(R)"           
#>  [73] "LAL(L)"            "AL-DC4(L)"         "AL-VM7d(L)"       
#>  [76] "CA(L)"             "b'L(L)"            "AL-D(R)"          
#>  [79] "b'2(R)"            "GNG"               "a3(R)"            
#>  [82] "AL-DA3(R)"         "NO3(R)"            "b2(R)"            
#>  [85] "AL-DA4m(L)"        "AL-VA5(R)"         "AL-VA6(R)"        
#>  [88] "WED(R)"            "FB-column3"        "PB(R3)"           
#>  [91] "AL(R)"             "AL-VM5d(R)"        "GOR(L)"           
#>  [94] "dACA(R)"           "EPA(R)"            "FBl9"             
#>  [97] "LOP(R)"            "FBl5"              "RUB(R)"           
#> [100] "AL-VL2p(R)"        "AOTU(R)"           "CRE(R)"           
#> [103] "AL-DA3(L)"         "CX"                "FB"               
#> [106] "NO2(R)"            "AOT(R)"            "EBr1"             
#> [109] "PB(R8)"            "gL(L)"             "AL(L)"            
#> [112] "GF(R)"             "mALT(R)"           "NO1(L)"           
#> [115] "ROB(R)"            "AL-VM3(R)"         "AL-DM2(L)"        
#> [118] "AL-VM5v(R)"        "aL(R)"             "FBl3"             
#> [121] "AL-DL5(R)"         "BU(L)"             "PB(L1)"           
#> [124] "NO3(L)"            "FBl6"              "EPA(L)"           
#> [127] "AL-DM6(R)"         "AL-VP4(R)"         "AL-VP1m(R)"       
#> [130] "NO2(L)"            "EB"                "CRE(-RUB)(L)"     
#> [133] "AL-DA4l(R)"        "MB(+ACA)(R)"       "AL-DA1(R)"        
#> [136] "EBr2r4"            "BU(R)"             "ICL(L)"           
#> [139] "b'1(R)"            "SPS(R)"            "MB(L)"            
#> [142] "PED(R)"            "ME(R)"             "PB(R6)"           
#> [145] "gL(R)"             "CRE(L)"            "AL-DM5(R)"        
#> [148] "AL-VP3(R)"         "vACA(R)"           "AMMC"             
#> [151] "g4(R)"             "SNP(L)"            "FBl7"             
#> [154] "g5(R)"             "AL-DM3(L)"         "AL-DM6(L)"        
#> [157] "AL-VL1(R)"         "a2(R)"             "hemibrain"        
#> [160] "GC"                "g3(R)"             "PB(L7)"           
#> [163] "LO(R)"             "AL-DC3(R)"         "a'2(R)"           
#> [166] "AVLP(R)"           "IPS(R)"            "POC"              
#> [169] "bL(L)"             "NO(L)"             "AL-DM2(R)"        
#> [172] "PB(L3)"            "SIP(L)"            "PB"               
#> [175] "SIP(R)"            "NO(R)"             "NO"               
#> [178] "AB(R)"             "AL-VC1(R)"         "AL-VM7v(L)"       
#> [181] "VES(L)"            "lACA(R)"           "AL-VA6(L)"        
#> [184] "bL(R)"             "AL-VA7l(R)"        "AL-VA1v(R)"       
#> [187] "AB(L)"             "a1(R)"             "SMP(L)"           
#> [190] "AL-DM1(L)"         "AL-DM5(L)"         "PB(R5)"           
#> [193] "SAD"               "LX(L)"             "RUB(L)"           
#> [196] "VES(R)"            "GA(R)"             "PB(R7)"           
#> [199] "SPS(L)"            "AL-DL3(R)"         "SCL(L)"           
#> [202] "AL-DC2(L)"         "a'1(R)"            "CAN(R)"           
#> [205] "LAL(-GA)(R)"       "FBl4"              "EBr6"             
#> [208] "a'L(R)"            "AL-VC3m(R)"        "FLA(R)"           
#> [211] "OL(R)"             "MB(R)"             "AL-DM4(L)"        
#> [214] "PB(R9)"            "AL-VL2a(R)"        "AL-VP5(R)"        
#> [217] "EBr3am"            "AL-DP1m(L)"        "FBl2"             
#> [220] "LH(R)"             "AL-VM1(R)"         "EBr5"             
#> [223] "AME(R)"            "AL-DC1(R)"         "AL-DA2(R)"        
#> [226] "AL-VA7m(R)"        "PB(L5)"            "PRW"              
#> [229] "g1(R)"             "PVLP(R)"          
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method  path                                        
#> [1,] "GET"   "/api/*"                                    
#> [2,] "POST"  "/api/v:ver/npexplorer/findneurons"         
#> [3,] "GET"   "/api/v:ver/raw/keyvalue/key/:instance/:key"
#> [4,] "GET"   "/api/v:ver/cached/roiconnectivity"         
#> [5,] "PATCH" "/api"                                      
#> [6,] "GET"   "/api/v:ver/npexplorer/commonconnectivity"  
#>      name                                                                
#> [1,] "github.com/labstack/echo.glob..func1"                              
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [5,] "github.com/labstack/echo.glob..func1"                              
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }