Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "EBr5"              "EB"                "LH(R)"            
#>   [4] "PB(L4)"            "PB(R7)"            "a'L(R)"           
#>   [7] "FBl4"              "ICL(R)"            "AOTU(R)"          
#>  [10] "EBr2r4"            "ROB(R)"            "OL(R)"            
#>  [13] "NO3(R)"            "b'L(L)"            "AL(R)"            
#>  [16] "AB(L)"             "AMMC"              "EBr6"             
#>  [19] "NO1(L)"            "EBr3d"             "VES(R)"           
#>  [22] "FBl7"              "a3(R)"             "LOP(R)"           
#>  [25] "GF(R)"             "b2(R)"             "NO1(R)"           
#>  [28] "PB(R8)"            "LX(R)"             "FBl3"             
#>  [31] "bL(R)"             "PB(L3)"            "INP"              
#>  [34] "b'1(R)"            "a'2(R)"            "SIP(L)"           
#>  [37] "mALT(R)"           "CA(R)"             "VES(L)"           
#>  [40] "PB(L5)"            "SMP(L)"            "b1(R)"            
#>  [43] "VMNP"              "SAD(-AMMC)"        "MB(+ACA)(R)"      
#>  [46] "vACA(R)"           "MB(L)"             "AVLP(R)"          
#>  [49] "a1(R)"             "CA(L)"             "SPS(R)"           
#>  [52] "PLP(R)"            "gL(R)"             "hemibrain"        
#>  [55] "CRE(R)"            "SIP(R)"            "PB(L7)"           
#>  [58] "FBl6"              "GA(R)"             "NO"               
#>  [61] "AOT(R)"            "a'3(R)"            "NO2(L)"           
#>  [64] "GC"                "SNP(R)"            "CAN(R)"           
#>  [67] "GNG"               "a'L(L)"            "VLNP(R)"          
#>  [70] "IB"                "CX"                "LAL(L)"           
#>  [73] "PVLP(R)"           "MB(R)"             "g1(R)"            
#>  [76] "PB(R3)"            "ATL(L)"            "PB(L9)"           
#>  [79] "PB(L6)"            "EBr3pw"            "SMP(R)"           
#>  [82] "EBr3am"            "PB(R9)"            "FLA(R)"           
#>  [85] "PB"                "PB(R1)"            "bL(L)"            
#>  [88] "AB(R)"             "RUB(R)"            "PRW"              
#>  [91] "BU(R)"             "LO(R)"             "PB(R5)"           
#>  [94] "lACA(R)"           "EPA(R)"            "NO(R)"            
#>  [97] "a'1(R)"            "FBl5"              "PB(R2)"           
#> [100] "PB(L8)"            "SLP(R)"            "CRE(L)"           
#> [103] "FBl1"              "FBl9"              "CRE(-ROB,-RUB)(R)"
#> [106] "GOR(L)"            "IPS(R)"            "g2(R)"            
#> [109] "PB(L1)"            "SNP(L)"            "aL(L)"            
#> [112] "LAL(R)"            "POC"               "WED(R)"           
#> [115] "SCL(R)"            "dACA(R)"           "SPS(L)"           
#> [118] "ATL(R)"            "EPA(L)"            "NO(L)"            
#> [121] "a2(R)"             "FB"                "b'L(R)"           
#> [124] "BU(L)"             "ME(R)"             "g4(R)"            
#> [127] "AME(R)"            "RUB(L)"            "PB(R6)"           
#> [130] "PB(L2)"            "LX(L)"             "FBl2"             
#> [133] "LAL(-GA)(R)"       "gL(L)"             "PED(R)"           
#> [136] "g3(R)"             "g5(R)"             "FBl8"             
#> [139] "PENP"              "EBr1"              "AL(L)"            
#> [142] "NO3(L)"            "ICL(L)"            "NO2(R)"           
#> [145] "GOR(R)"            "SCL(L)"            "b'2(R)"           
#> [148] "mALT(L)"           "aL(R)"             "PB(R4)"           
#> [151] "SAD"              
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method   path                                
#> [1,] "POST"   "/logout"                           
#> [2,] "GET"    "/profile"                          
#> [3,] "GET"    "/api"                              
#> [4,] "REPORT" "/api"                              
#> [5,] "POST"   "/api/npexplorer/roisinneuron"      
#> [6,] "GET"    "/api/v:ver/npexplorer/distribution"
#>      name                                                                                        
#> [1,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [2,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [3,] "github.com/labstack/echo.glob..func1"                                                      
#> [4,] "github.com/labstack/echo.glob..func1"                                                      
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }