Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "LX(L)"             "bL(R)"             "AL-VA4(R)"        
#>   [4] "AL-VA7m(R)"        "GOR(L)"            "AL-DM1(L)"        
#>   [7] "a'1(R)"            "b'1(R)"            "gL(R)"            
#>  [10] "EBr6"              "VES(R)"            "GA(R)"            
#>  [13] "AL-VC5(R)"         "AL-VM7v(R)"        "FBl6"             
#>  [16] "AL-VC1(R)"         "AL-DM6(L)"         "a1(R)"            
#>  [19] "BU(L)"             "CRE(R)"            "NO1(R)"           
#>  [22] "AL-DM4(L)"         "AB(R)"             "b'L(L)"           
#>  [25] "AL-DA4l(R)"        "AL-VM7d(R)"        "IB"               
#>  [28] "SAD"               "FBl9"              "GNG"              
#>  [31] "FBl2"              "EBr3am"            "AL-DL3(R)"        
#>  [34] "AL-VA5(R)"         "AMMC"              "EBr3d"            
#>  [37] "PB(L4)"            "MB(R)"             "AL-DA4m(R)"       
#>  [40] "AL-DC2(R)"         "AB(L)"             "IPS(R)"           
#>  [43] "EBr5"              "AL-DC1(R)"         "a'3(R)"           
#>  [46] "g3(R)"             "AL-DA3(L)"         "FBl3"             
#>  [49] "SMP(R)"            "PRW"               "SNP(L)"           
#>  [52] "IB112"             "LAL(R)"            "AL-DC2(L)"        
#>  [55] "b'2(R)"            "g1(R)"             "LH(R)"            
#>  [58] "RUB(L)"            "SCL(L)"            "PB(R6)"           
#>  [61] "OL(R)"             "AL-DA1(R)"         "AL-DM6(R)"        
#>  [64] "ICL(L)"            "PB(R4)"            "AL(R)"            
#>  [67] "AL-VM5v(R)"        "b'L(R)"            "CRE(-RUB)(L)"     
#>  [70] "SLP(R)"            "gL(L)"             "a2(R)"            
#>  [73] "AL-VL2p(R)"        "PB(R7)"            "AOTU(R)"          
#>  [76] "LAL(-GA)(R)"       "ME(R)"             "AL-VP1d(R)"       
#>  [79] "LAL(L)"            "AL-VP2(R)"         "AOT(R)"           
#>  [82] "AL-VM1(R)"         "AL-VL1(R)"         "AL-VP4(R)"        
#>  [85] "SIP(R)"            "AL-VA1d(R)"        "NO(L)"            
#>  [88] "ROB(R)"            "AL-DA3(R)"         "g2(R)"            
#>  [91] "AL-DC4(R)"         "CRE(-ROB,-RUB)(R)" "AL-DM2(L)"        
#>  [94] "NO3(R)"            "g4(R)"             "AL-D(L)"          
#>  [97] "FBl4"              "AL-D(R)"           "lACA(R)"          
#> [100] "aL(R)"             "GC"                "AL-DC1(L)"        
#> [103] "AL-DC4(L)"         "AL-DL5(L)"         "AL-DM3(L)"        
#> [106] "PB(R9)"            "PB(L9)"            "CRE(L)"           
#> [109] "SMP(L)"            "vACA(R)"           "AL-DL4(R)"        
#> [112] "VLNP(R)"           "PB(L1)"            "AL(L)"            
#> [115] "EBr3pw"            "AL-DM1(R)"         "AL-VP1l(R)"       
#> [118] "SPS(R)"            "VMNP"              "NO"               
#> [121] "PB"                "FBl7"              "AL-VP1m(R)"       
#> [124] "ATL(L)"            "a'2(R)"            "EB"               
#> [127] "AL-DL4(L)"         "AL-DM5(L)"         "SCL(R)"           
#> [130] "SNP(R)"            "AL-VA6(L)"         "FLA(R)"           
#> [133] "LX(R)"             "ATL(R)"            "a'L(L)"           
#> [136] "AL-DA2(R)"         "AL-VC2(R)"         "mALT(R)"          
#> [139] "AL-VL2a(R)"        "AL-DM3(R)"         "SPS(L)"           
#> [142] "NO2(L)"            "EBr1"              "AL-DP1m(L)"       
#> [145] "b1(R)"             "AL-DL2d(R)"        "POC"              
#> [148] "dACA(R)"           "AL-VM7d(L)"        "MB(+ACA)(R)"      
#> [151] "AL-VM3(R)"         "PB(R5)"            "PLP(R)"           
#> [154] "FBl1"              "AL-VA7l(R)"        "LOP(R)"           
#> [157] "CX"                "PB(R1)"            "NO3(L)"           
#> [160] "SAD(-AMMC)"        "VES(L)"            "AL-VP3(R)"        
#> [163] "NO1(L)"            "AL-VM7v(L)"        "a3(R)"            
#> [166] "MB(L)"             "AL-DP1l(R)"        "AL-VA3(R)"        
#> [169] "INP"               "PB(R8)"            "AL-DA2(L)"        
#> [172] "FBl8"              "PB(R2)"            "PB(R3)"           
#> [175] "AL-DM5(R)"         "WED(R)"            "AME(R)"           
#> [178] "AL-DP1m(R)"        "AL-VC3m(R)"        "LO(R)"            
#> [181] "AL-DL1(R)"         "PB(L3)"            "AL-VA1v(R)"       
#> [184] "FB"                "CAN(R)"            "AL-VA2(R)"        
#> [187] "RUB(R)"            "CA(L)"             "FB-column3"       
#> [190] "GOR(R)"            "AL-VM4(R)"         "AL-VP5(R)"        
#> [193] "PB(L8)"            "SIP(L)"            "EPA(R)"           
#> [196] "bL(L)"             "AL-VM5d(R)"        "PENP"             
#> [199] "AL-DA4m(L)"        "PVLP(R)"           "FBl5"             
#> [202] "CA(R)"             "aL(L)"             "EPA(L)"           
#> [205] "AL-DM4(R)"         "b2(R)"             "AL-DM2(R)"        
#> [208] "AL-VC3l(R)"        "PED(R)"            "BU(R)"            
#> [211] "AL-DL2v(R)"        "PB(L5)"            "EBr2r4"           
#> [214] "AVLP(R)"           "NO(R)"             "g5(R)"            
#> [217] "AL-DC3(R)"         "AL-V(R)"           "PB(L2)"           
#> [220] "a'L(R)"            "AL-DL5(R)"         "AL-VM2(R)"        
#> [223] "PB(L7)"            "AL-VC4(R)"         "PB(L6)"           
#> [226] "hemibrain"         "ICL(R)"            "NO2(R)"           
#> [229] "AL-VA6(R)"         "mALT(L)"           "GF(R)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "a'L(L)"            "GC"                "aL(R)"            
#>   [4] "EB"                "CRE(L)"            "FBl6"             
#>   [7] "GOR(R)"            "NO(L)"             "b'L(R)"           
#>  [10] "a'L(R)"            "PB(R5)"            "LOP(R)"           
#>  [13] "b2(R)"             "SPS(L)"            "a'1(R)"           
#>  [16] "bL(L)"             "FBl8"              "NO2(R)"           
#>  [19] "g2(R)"             "NO"                "RUB(L)"           
#>  [22] "VLNP(R)"           "CRE(-ROB,-RUB)(R)" "PB(R9)"           
#>  [25] "ROB(R)"            "AB(R)"             "b'1(R)"           
#>  [28] "b'2(R)"            "mALT(R)"           "PB(L8)"           
#>  [31] "INP"               "IB"                "SMP(L)"           
#>  [34] "a2(R)"             "SPS(R)"            "FB"               
#>  [37] "LH(R)"             "CA(L)"             "CRE(R)"           
#>  [40] "AL(L)"             "PB(L6)"            "CAN(R)"           
#>  [43] "PB(L7)"            "MB(+ACA)(R)"       "PB(L5)"           
#>  [46] "g3(R)"             "a'2(R)"            "NO(R)"            
#>  [49] "FBl7"              "WED(R)"            "LX(R)"            
#>  [52] "EBr3pw"            "FBl1"              "vACA(R)"          
#>  [55] "PB(R6)"            "NO3(R)"            "PB(L9)"           
#>  [58] "VES(R)"            "ATL(R)"            "NO2(L)"           
#>  [61] "LAL(L)"            "EBr3d"             "EBr2r4"           
#>  [64] "BU(R)"             "PB(R3)"            "b1(R)"            
#>  [67] "PB(R4)"            "SMP(R)"            "AOTU(R)"          
#>  [70] "lACA(R)"           "EBr1"              "NO1(R)"           
#>  [73] "LX(L)"             "LAL(R)"            "AOT(R)"           
#>  [76] "PED(R)"            "dACA(R)"           "PLP(R)"           
#>  [79] "aL(L)"             "PB(L2)"            "hemibrain"        
#>  [82] "PB(R7)"            "SNP(R)"            "EPA(L)"           
#>  [85] "MB(R)"             "gL(L)"             "SCL(L)"           
#>  [88] "b'L(L)"            "GOR(L)"            "PB(R8)"           
#>  [91] "SIP(R)"            "SLP(R)"            "mALT(L)"          
#>  [94] "OL(R)"             "AMMC"              "SIP(L)"           
#>  [97] "gL(R)"             "FBl4"              "AVLP(R)"          
#> [100] "bL(R)"             "GA(R)"             "SCL(R)"           
#> [103] "RUB(R)"            "g1(R)"             "BU(L)"            
#> [106] "PB(L4)"            "CX"                "PB(L1)"           
#> [109] "g4(R)"             "SAD(-AMMC)"        "ICL(R)"           
#> [112] "ATL(L)"            "ME(R)"             "GNG"              
#> [115] "POC"               "EBr3am"            "PENP"             
#> [118] "EBr6"              "PB"                "NO3(L)"           
#> [121] "a'3(R)"            "PB(L3)"            "NO1(L)"           
#> [124] "FLA(R)"            "SNP(L)"            "CA(R)"            
#> [127] "a1(R)"             "GF(R)"             "SAD"              
#> [130] "EPA(R)"            "ICL(L)"            "g5(R)"            
#> [133] "FBl3"              "PRW"               "MB(L)"            
#> [136] "PB(R1)"            "LO(R)"             "AL(R)"            
#> [139] "AB(L)"             "FBl2"              "FBl5"             
#> [142] "FBl9"              "a3(R)"             "VES(L)"           
#> [145] "LAL(-GA)(R)"       "EBr5"              "VMNP"             
#> [148] "PVLP(R)"           "PB(R2)"            "IPS(R)"           
#> [151] "AME(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DL3(R)"         "AL-DL5(R)"         "PENP"             
#>   [4] "mALT(R)"           "SCL(R)"            "AL-VM5v(R)"       
#>   [7] "CRE(L)"            "LAL(L)"            "PED(R)"           
#>  [10] "INP"               "FBl3"              "b1(R)"            
#>  [13] "AL(R)"             "SIP(R)"            "FB"               
#>  [16] "PB(R8)"            "PRW"               "g2(R)"            
#>  [19] "AL-VL2a(R)"        "AL-VM5d(R)"        "NO(R)"            
#>  [22] "PB(R3)"            "OL(R)"             "EBr1"             
#>  [25] "GNG"               "mALT(L)"           "a1(R)"            
#>  [28] "LAL(-GA)(R)"       "AL-DC4(L)"         "GOR(R)"           
#>  [31] "PB(L4)"            "AL-VA1d(R)"        "AL-DL4(L)"        
#>  [34] "AL-VM7d(L)"        "EBr3pw"            "a'L(L)"           
#>  [37] "AL-DP1m(L)"        "a'1(R)"            "CA(R)"            
#>  [40] "RUB(L)"            "AL-DA4m(L)"        "AL-VP1d(R)"       
#>  [43] "VES(L)"            "AL-DA3(R)"         "PB(R6)"           
#>  [46] "SIP(L)"            "EBr5"              "AL-D(L)"          
#>  [49] "AL-DM4(L)"         "AL-DA2(L)"         "AL-VC2(R)"        
#>  [52] "PB(R5)"            "WED(R)"            "NO1(R)"           
#>  [55] "AL-D(R)"           "AL-VM4(R)"         "IB"               
#>  [58] "GC"                "dACA(R)"           "AL-VA6(L)"        
#>  [61] "AL-DA4m(R)"        "AL-DL2v(R)"        "NO2(R)"           
#>  [64] "AL-VA5(R)"         "PB(L5)"            "SAD(-AMMC)"       
#>  [67] "AL-VP1l(R)"        "SCL(L)"            "PLP(R)"           
#>  [70] "AL-DL5(L)"         "NO3(L)"            "LH(R)"            
#>  [73] "AL-DC4(R)"         "SAD"               "AME(R)"           
#>  [76] "bL(R)"             "AL-VC5(R)"         "AL-VP2(R)"        
#>  [79] "LX(R)"             "EPA(L)"            "ATL(L)"           
#>  [82] "AL-VM7d(R)"        "RUB(R)"            "AL-VA3(R)"        
#>  [85] "aL(R)"             "SNP(R)"            "AVLP(R)"          
#>  [88] "AL-V(R)"           "AOTU(R)"           "NO3(R)"           
#>  [91] "g4(R)"             "AL(L)"             "PB(L9)"           
#>  [94] "SNP(L)"            "PB(L3)"            "AL-DP1m(R)"       
#>  [97] "MB(R)"             "b2(R)"             "ROB(R)"           
#> [100] "PB(R2)"            "aL(L)"             "NO(L)"            
#> [103] "GOR(L)"            "IPS(R)"            "LAL(R)"           
#> [106] "AL-VC1(R)"         "CA(L)"             "CRE(R)"           
#> [109] "ME(R)"             "VES(R)"            "AL-VM3(R)"        
#> [112] "CRE(-ROB,-RUB)(R)" "FBl7"              "FBl8"             
#> [115] "AL-VA7l(R)"        "GF(R)"             "GA(R)"            
#> [118] "a'2(R)"            "ICL(L)"            "LO(R)"            
#> [121] "AB(L)"             "gL(L)"             "bL(L)"            
#> [124] "LX(L)"             "AL-VA4(R)"         "AL-VM1(R)"        
#> [127] "AL-VL1(R)"         "PB"                "EBr2r4"           
#> [130] "AL-DC2(L)"         "FB-column3"        "FBl4"             
#> [133] "g1(R)"             "AL-DM1(L)"         "NO"               
#> [136] "AL-VC4(R)"         "g3(R)"             "AL-DM6(L)"        
#> [139] "CAN(R)"            "CX"                "AL-VA6(R)"        
#> [142] "NO1(L)"            "BU(L)"             "AL-DA3(L)"        
#> [145] "PB(R1)"            "FBl9"              "EBr6"             
#> [148] "b'L(L)"            "PB(L6)"            "FBl1"             
#> [151] "AL-DL2d(R)"        "AL-DM1(R)"         "b'2(R)"           
#> [154] "PB(L8)"            "SPS(L)"            "AL-DA4l(R)"       
#> [157] "AL-VP4(R)"         "CRE(-RUB)(L)"      "PB(L7)"           
#> [160] "SLP(R)"            "AL-DM3(R)"         "AMMC"             
#> [163] "AL-DM2(R)"         "LOP(R)"            "AL-VM7v(R)"       
#> [166] "hemibrain"         "a'L(R)"            "FBl5"             
#> [169] "AL-DP1l(R)"        "AL-VA1v(R)"        "NO2(L)"           
#> [172] "SPS(R)"            "AL-VC3l(R)"        "AL-VP5(R)"        
#> [175] "b'1(R)"            "vACA(R)"           "b'L(R)"           
#> [178] "VMNP"              "AL-VP1m(R)"        "AL-DC1(R)"        
#> [181] "AOT(R)"            "SMP(R)"            "ATL(R)"           
#> [184] "ICL(R)"            "PB(R4)"            "AL-DC2(R)"        
#> [187] "AL-VA7m(R)"        "EB"                "EPA(R)"           
#> [190] "AL-DL1(R)"         "AL-DM3(L)"         "PB(R9)"           
#> [193] "g5(R)"             "gL(R)"             "AB(R)"            
#> [196] "AL-VC3m(R)"        "POC"               "AL-DM5(R)"        
#> [199] "AL-DM6(R)"         "AL-VM2(R)"         "PB(L2)"           
#> [202] "EBr3am"            "AL-DM5(L)"         "PVLP(R)"          
#> [205] "AL-VP3(R)"         "a2(R)"             "AL-DA1(R)"        
#> [208] "AL-DL4(R)"         "VLNP(R)"           "FLA(R)"           
#> [211] "AL-DC1(L)"         "AL-DA2(R)"         "AL-DC3(R)"        
#> [214] "AL-DM4(R)"         "BU(R)"             "AL-DM2(L)"        
#> [217] "EBr3d"             "a'3(R)"            "PB(L1)"           
#> [220] "SMP(L)"            "lACA(R)"           "AL-VL2p(R)"       
#> [223] "FBl6"              "a3(R)"             "AL-VA2(R)"        
#> [226] "PB(R7)"            "MB(L)"             "AL-VM7v(L)"       
#> [229] "MB(+ACA)(R)"       "FBl2"             
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AB(L)"             "PB(R6)"            "GA(R)"            
#>   [4] "ATL(R)"            "AL-DM4(R)"         "MB(+ACA)(R)"      
#>   [7] "AL-DC3(R)"         "AL-DM3(R)"         "AL-VP3(R)"        
#>  [10] "FBl6"              "AL-VC4(R)"         "AL-DA4l(R)"       
#>  [13] "AL-VA1d(R)"        "AL-VA6(R)"         "AL-DA4m(L)"       
#>  [16] "PB(L7)"            "PB(R4)"            "AL-VA2(R)"        
#>  [19] "SNP(R)"            "LO(R)"             "AL-VA4(R)"        
#>  [22] "AOTU(R)"           "BU(L)"             "EBr3d"            
#>  [25] "NO"                "mALT(R)"           "FBl4"             
#>  [28] "AL-DM6(R)"         "ATL(L)"            "ICL(L)"           
#>  [31] "FBl9"              "AL-VA5(R)"         "SMP(R)"           
#>  [34] "g1(R)"             "FLA(R)"            "SCL(L)"           
#>  [37] "PLP(R)"            "PB(R3)"            "a'1(R)"           
#>  [40] "CAN(R)"            "AL-DA1(R)"         "AL-VM3(R)"        
#>  [43] "AL-DL5(R)"         "AL-VM1(R)"         "PB"               
#>  [46] "SAD(-AMMC)"        "AL-DL2d(R)"        "AL-DL2v(R)"       
#>  [49] "GOR(R)"            "AL-DL3(R)"         "AL-VA7l(R)"       
#>  [52] "LX(R)"             "a'L(R)"            "FBl7"             
#>  [55] "hemibrain"         "BU(R)"             "CRE(-RUB)(L)"     
#>  [58] "AL-VM5v(R)"        "FB"                "EBr6"             
#>  [61] "AL-DA3(L)"         "AL-DM5(R)"         "AL-VL2a(R)"       
#>  [64] "VLNP(R)"           "AL-DA2(L)"         "aL(L)"            
#>  [67] "SLP(R)"            "AL-VL1(R)"         "AL-VC5(R)"        
#>  [70] "RUB(R)"            "AL-DM1(L)"         "PVLP(R)"          
#>  [73] "NO(L)"             "AL-DC1(R)"         "GOR(L)"           
#>  [76] "SCL(R)"            "PB(R8)"            "EBr1"             
#>  [79] "AL-DC4(R)"         "AL-VA3(R)"         "AL-VM7v(R)"       
#>  [82] "AL-DC2(L)"         "AL-DM5(L)"         "AL-DA2(R)"        
#>  [85] "AL-VA6(L)"         "NO3(L)"            "AL-DA3(R)"        
#>  [88] "lACA(R)"           "PB(R9)"            "LX(L)"            
#>  [91] "a'L(L)"            "AL-DM3(L)"         "MB(R)"            
#>  [94] "EBr2r4"            "AL-DM6(L)"         "AL-VM7d(L)"       
#>  [97] "a1(R)"             "SAD"               "EBr5"             
#> [100] "AL-DP1m(R)"        "AL-VA1v(R)"        "NO1(L)"           
#> [103] "a'3(R)"            "PB(R5)"            "b'1(R)"           
#> [106] "NO2(L)"            "aL(R)"             "ME(R)"            
#> [109] "FBl2"              "CRE(-ROB,-RUB)(R)" "AL-VM7v(L)"       
#> [112] "bL(R)"             "b'2(R)"            "FB-column3"       
#> [115] "PB(R2)"            "AL-VP2(R)"         "PB(L9)"           
#> [118] "LOP(R)"            "a2(R)"             "AL-VC2(R)"        
#> [121] "CX"                "PB(L1)"            "CRE(L)"           
#> [124] "VES(R)"            "AL-VM2(R)"         "AL-DM2(L)"        
#> [127] "RUB(L)"            "FBl1"              "AOT(R)"           
#> [130] "IB"                "AL-DC4(L)"         "bL(L)"            
#> [133] "AL-DM2(R)"         "AL-V(R)"           "NO1(R)"           
#> [136] "SMP(L)"            "PED(R)"            "PB(L6)"           
#> [139] "mALT(L)"           "GC"                "LH(R)"            
#> [142] "vACA(R)"           "GF(R)"             "AME(R)"           
#> [145] "AL-VA7m(R)"        "AL-VP4(R)"         "SPS(R)"           
#> [148] "a3(R)"             "AL-VP5(R)"         "PB(L2)"           
#> [151] "AL-D(R)"           "AL-D(L)"           "SIP(R)"           
#> [154] "EBr3pw"            "OL(R)"             "AVLP(R)"          
#> [157] "g2(R)"             "MB(L)"             "AL-VM4(R)"        
#> [160] "PB(L5)"            "FBl3"              "AB(R)"            
#> [163] "IPS(R)"            "AL-VC3l(R)"        "PB(R7)"           
#> [166] "dACA(R)"           "EB"                "AL-DM4(L)"        
#> [169] "POC"               "g3(R)"             "PRW"              
#> [172] "SIP(L)"            "AL-DL4(L)"         "AL-DL1(R)"        
#> [175] "AL-DP1m(L)"        "ROB(R)"            "CA(R)"            
#> [178] "VES(L)"            "CRE(R)"            "PB(L8)"           
#> [181] "AL-DC2(R)"         "EPA(R)"            "ICL(R)"           
#> [184] "AL-VL2p(R)"        "PB(L3)"            "FBl5"             
#> [187] "b1(R)"             "PENP"              "AL-VC3m(R)"       
#> [190] "AL-VP1d(R)"        "gL(R)"             "EPA(L)"           
#> [193] "GNG"               "AL-DL4(R)"         "LAL(L)"           
#> [196] "g4(R)"             "AL-DL5(L)"         "NO(R)"            
#> [199] "g5(R)"             "PB(L4)"            "AL-VM7d(R)"       
#> [202] "LAL(-GA)(R)"       "a'2(R)"            "AL-DC1(L)"        
#> [205] "INP"               "b2(R)"             "gL(L)"            
#> [208] "AMMC"              "CA(L)"             "b'L(R)"           
#> [211] "FBl8"              "AL-VM5d(R)"        "EBr3am"           
#> [214] "SPS(L)"            "VMNP"              "AL(L)"            
#> [217] "LAL(R)"            "NO3(R)"            "NO2(R)"           
#> [220] "b'L(L)"            "AL-DA4m(R)"        "AL-VP1m(R)"       
#> [223] "AL-VP1l(R)"        "WED(R)"            "AL-DM1(R)"        
#> [226] "AL-DP1l(R)"        "AL-VC1(R)"         "PB(R1)"           
#> [229] "SNP(L)"            "AL(R)"            
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method     path                               
#> [1,] "POST"     "/api/v:ver/npexplorer/rankedtable"
#> [2,] "PATCH"    "/api"                             
#> [3,] "PROPFIND" "/api/*"                           
#> [4,] "GET"      "/*"                               
#> [5,] "GET"      "/api/v:ver/cached/roiconnectivity"
#> [6,] "HEAD"     "/api/*"                           
#>      name                                                                
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [2,] "github.com/labstack/echo.glob..func1"                              
#> [3,] "github.com/labstack/echo.glob..func1"                              
#> [4,] "github.com/labstack/echo.common.static.func1"                      
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [6,] "github.com/labstack/echo.glob..func1"                              

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }