Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $fib19
#> $fib19$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $fib19$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $fib19$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $fib19$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $fib19$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $fib19$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "dACA(R)"           "PB(L7)"            "PRW"              
#>   [4] "ME(R)"             "mALT(L)"           "AL-VP1m(R)"       
#>   [7] "SPS(L)"            "AL-DL4(L)"         "AL-DL3(R)"        
#>  [10] "AL-VA2(R)"         "AL-DA3(L)"         "IB112"            
#>  [13] "SLP(R)"            "ICL(L)"            "CA(R)"            
#>  [16] "AL-VC2(R)"         "AL-VC3m(R)"        "lACA(R)"          
#>  [19] "CRE(R)"            "PB(R9)"            "PB(L2)"           
#>  [22] "PENP"              "PLP(R)"            "AL-VC4(R)"        
#>  [25] "ROB(R)"            "EBr3pw"            "AL-VP3(R)"        
#>  [28] "GA(R)"             "g4(R)"             "CA(L)"            
#>  [31] "EBr5"              "AL(L)"             "AL-DC2(L)"        
#>  [34] "NO"                "AL-DL2v(R)"        "AL-VA7m(R)"       
#>  [37] "a'L(R)"            "AL-D(L)"           "AL-DL2d(R)"       
#>  [40] "LAL(-GA)(R)"       "CAN(R)"            "RUB(L)"           
#>  [43] "AL-DC4(R)"         "AL-VP4(R)"         "AL-DM1(L)"        
#>  [46] "FBl2"              "AME(R)"            "PB(R4)"           
#>  [49] "AL-DM1(R)"         "AL-DA2(L)"         "FBl1"             
#>  [52] "LX(L)"             "PED(R)"            "AL-D(R)"          
#>  [55] "AL-VP1d(R)"        "b'1(R)"            "FB"               
#>  [58] "b1(R)"             "AL-VM1(R)"         "NO1(R)"           
#>  [61] "PB(L9)"            "AL-VM3(R)"         "SPS(R)"           
#>  [64] "SCL(R)"            "EB"                "PB"               
#>  [67] "VMNP"              "EPA(R)"            "SNP(L)"           
#>  [70] "PVLP(R)"           "AL-DP1m(R)"        "AL-VA1d(R)"       
#>  [73] "AL-VC5(R)"         "b'L(R)"            "AL-DM4(L)"        
#>  [76] "g5(R)"             "gL(R)"             "AL-DA3(R)"        
#>  [79] "LOP(R)"            "bL(R)"             "AL-VP2(R)"        
#>  [82] "ATL(L)"            "CX"                "NO2(R)"           
#>  [85] "b'L(L)"            "AL-VA7l(R)"        "EBr3d"            
#>  [88] "AL-VA5(R)"         "AL-VM5d(R)"        "IB"               
#>  [91] "AOTU(R)"           "PB(R3)"            "LAL(R)"           
#>  [94] "AB(L)"             "LO(R)"             "FBl7"             
#>  [97] "AL(R)"             "GF(R)"             "AL-VA3(R)"        
#> [100] "a'1(R)"            "NO2(L)"            "AL-DC2(R)"        
#> [103] "AL-DM4(R)"         "AL-VP5(R)"         "AL-V(R)"          
#> [106] "EBr6"              "INP"               "AL-VL2a(R)"       
#> [109] "GOR(R)"            "AL-VM2(R)"         "SMP(R)"           
#> [112] "aL(R)"             "NO(R)"             "NO(L)"            
#> [115] "a'3(R)"            "AL-DC4(L)"         "AL-DL1(R)"        
#> [118] "AL-VP1l(R)"        "PB(R2)"            "AL-VA4(R)"        
#> [121] "AL-VA6(R)"         "EBr3am"            "NO1(L)"           
#> [124] "AL-DM3(L)"         "MB(+ACA)(R)"       "SIP(R)"           
#> [127] "SAD(-AMMC)"        "AL-VC3l(R)"        "MB(L)"            
#> [130] "AL-DA4m(R)"        "RUB(R)"            "AL-DM5(L)"        
#> [133] "bL(L)"             "a'L(L)"            "AL-DL5(R)"        
#> [136] "AL-VM5v(R)"        "FLA(R)"            "PB(R8)"           
#> [139] "gL(L)"             "ATL(R)"            "AL-VM7d(R)"       
#> [142] "SAD"               "PB(L8)"            "AL-VM7d(L)"       
#> [145] "PB(L3)"            "NO3(L)"            "SCL(L)"           
#> [148] "AL-VL2p(R)"        "AMMC"              "ICL(R)"           
#> [151] "OL(R)"             "VES(L)"            "AL-VC1(R)"        
#> [154] "POC"               "LX(R)"             "SNP(R)"           
#> [157] "VES(R)"            "CRE(-ROB,-RUB)(R)" "AL-DM2(L)"        
#> [160] "FBl4"              "PB(L5)"            "aL(L)"            
#> [163] "a3(R)"             "AL-DA4l(R)"        "g1(R)"            
#> [166] "NO3(R)"            "BU(L)"             "AL-VL1(R)"        
#> [169] "g3(R)"             "LH(R)"             "AL-DC3(R)"        
#> [172] "FBl8"              "b'2(R)"            "PB(L1)"           
#> [175] "AL-DA4m(L)"        "PB(R7)"            "PB(L4)"           
#> [178] "VLNP(R)"           "FBl5"              "AL-DP1l(R)"       
#> [181] "SMP(L)"            "AL-DM6(L)"         "PB(L6)"           
#> [184] "a'2(R)"            "EBr2r4"            "PB(R1)"           
#> [187] "EPA(L)"            "AL-DA2(R)"         "AOT(R)"           
#> [190] "PB(R5)"            "AL-DP1m(L)"        "CRE(-RUB)(L)"     
#> [193] "FB-column3"        "AL-VM7v(R)"        "hemibrain"        
#> [196] "AL-VA6(L)"         "AL-VM7v(L)"        "LAL(L)"           
#> [199] "IPS(R)"            "AL-DM5(R)"         "FBl9"             
#> [202] "GOR(L)"            "WED(R)"            "BU(R)"            
#> [205] "FBl3"              "AB(R)"             "AL-DA1(R)"        
#> [208] "AL-DC1(R)"         "AL-DL4(R)"         "PB(R6)"           
#> [211] "EBr1"              "CRE(L)"            "a1(R)"            
#> [214] "a2(R)"             "g2(R)"             "AL-DM3(R)"        
#> [217] "FBl6"              "MB(R)"             "AL-DM6(R)"        
#> [220] "AL-VM4(R)"         "SIP(L)"            "GC"               
#> [223] "GNG"               "AVLP(R)"           "AL-DM2(R)"        
#> [226] "mALT(R)"           "AL-DC1(L)"         "AL-DL5(L)"        
#> [229] "b2(R)"             "AL-VA1v(R)"        "vACA(R)"          
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(R8)"            "a'2(R)"            "ME(R)"            
#>   [4] "GA(R)"             "LAL(-GA)(R)"       "bL(L)"            
#>   [7] "EBr3pw"            "LAL(R)"            "CRE(-ROB,-RUB)(R)"
#>  [10] "a'3(R)"            "g3(R)"             "EBr3d"            
#>  [13] "FBl6"              "OL(R)"             "b'1(R)"           
#>  [16] "LOP(R)"            "mALT(L)"           "SPS(R)"           
#>  [19] "gL(L)"             "CA(R)"             "b1(R)"            
#>  [22] "AB(R)"             "GOR(R)"            "AOTU(R)"          
#>  [25] "PB(R7)"            "SPS(L)"            "AL(L)"            
#>  [28] "CRE(L)"            "EBr6"              "LX(R)"            
#>  [31] "NO2(R)"            "SLP(R)"            "PB(R6)"           
#>  [34] "PB"                "FBl5"              "PB(L2)"           
#>  [37] "NO1(L)"            "SIP(L)"            "PED(R)"           
#>  [40] "ICL(L)"            "AOT(R)"            "b'2(R)"           
#>  [43] "dACA(R)"           "NO"                "VES(L)"           
#>  [46] "SAD"               "PB(R4)"            "b'L(R)"           
#>  [49] "PB(L8)"            "EBr2r4"            "CRE(R)"           
#>  [52] "FBl1"              "FBl2"              "LH(R)"            
#>  [55] "EBr3am"            "PB(L9)"            "bL(R)"            
#>  [58] "FBl4"              "FBl7"              "GNG"              
#>  [61] "IB"                "a'L(R)"            "SIP(R)"           
#>  [64] "LO(R)"             "hemibrain"         "EB"               
#>  [67] "ATL(R)"            "MB(L)"             "VES(R)"           
#>  [70] "PB(R3)"            "AVLP(R)"           "RUB(L)"           
#>  [73] "NO1(R)"            "SMP(R)"            "PB(L5)"           
#>  [76] "NO2(L)"            "PB(L1)"            "PLP(R)"           
#>  [79] "g4(R)"             "AME(R)"            "INP"              
#>  [82] "PB(R5)"            "PENP"              "lACA(R)"          
#>  [85] "SNP(L)"            "aL(L)"             "CX"               
#>  [88] "gL(R)"             "EPA(L)"            "AB(L)"            
#>  [91] "vACA(R)"           "LX(L)"             "PB(L3)"           
#>  [94] "PRW"               "EBr5"              "FLA(R)"           
#>  [97] "EPA(R)"            "NO3(L)"            "GF(R)"            
#> [100] "g1(R)"             "g2(R)"             "BU(R)"            
#> [103] "SAD(-AMMC)"        "a'L(L)"            "MB(+ACA)(R)"      
#> [106] "VLNP(R)"           "ATL(L)"            "AMMC"             
#> [109] "IPS(R)"            "CAN(R)"            "GC"               
#> [112] "NO3(R)"            "FB"                "EBr1"             
#> [115] "SNP(R)"            "PB(L7)"            "POC"              
#> [118] "a'1(R)"            "FBl8"              "b'L(L)"           
#> [121] "NO(L)"             "PB(L4)"            "aL(R)"            
#> [124] "BU(L)"             "PB(R2)"            "RUB(R)"           
#> [127] "a2(R)"             "FBl9"              "ICL(R)"           
#> [130] "AL(R)"             "a1(R)"             "NO(R)"            
#> [133] "g5(R)"             "PVLP(R)"           "FBl3"             
#> [136] "MB(R)"             "SMP(L)"            "GOR(L)"           
#> [139] "SCL(R)"            "ROB(R)"            "PB(L6)"           
#> [142] "PB(R1)"            "a3(R)"             "LAL(L)"           
#> [145] "mALT(R)"           "WED(R)"            "b2(R)"            
#> [148] "VMNP"              "PB(R9)"            "SCL(L)"           
#> [151] "CA(L)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "NO"                "SLP(R)"            "AL-DA2(R)"        
#>   [4] "PB(R4)"            "PB(R5)"            "GC"               
#>   [7] "NO1(L)"            "FBl1"              "AVLP(R)"          
#>  [10] "AL-VC5(R)"         "AL-V(R)"           "g2(R)"            
#>  [13] "PB(R1)"            "a'L(L)"            "AL-VA7l(R)"       
#>  [16] "SMP(L)"            "b'L(R)"            "g1(R)"            
#>  [19] "EPA(R)"            "AL-VM2(R)"         "BU(L)"            
#>  [22] "PB"                "PB(R9)"            "AL-VM1(R)"        
#>  [25] "LAL(L)"            "hemibrain"         "SPS(R)"           
#>  [28] "LX(R)"             "AL-VA2(R)"         "LX(L)"            
#>  [31] "AL-DC2(R)"         "AL-DM4(R)"         "AL-VA1d(R)"       
#>  [34] "AL-DL2d(R)"        "AL-VA1v(R)"        "a'3(R)"           
#>  [37] "GF(R)"             "SAD"               "aL(R)"            
#>  [40] "AL-VM7d(L)"        "AL-DA2(L)"         "VES(L)"           
#>  [43] "RUB(R)"            "lACA(R)"           "EB"               
#>  [46] "bL(L)"             "AL-DC3(R)"         "NO(L)"            
#>  [49] "AL-VP3(R)"         "PB(L5)"            "AL-DM2(L)"        
#>  [52] "SAD(-AMMC)"        "FBl3"              "INP"              
#>  [55] "AL-DM3(L)"         "a'1(R)"            "POC"              
#>  [58] "VMNP"              "PB(R2)"            "EPA(L)"           
#>  [61] "RUB(L)"            "AL-VA4(R)"         "LAL(-GA)(R)"      
#>  [64] "AL-D(L)"           "gL(R)"             "FBl4"             
#>  [67] "AL-DP1m(R)"        "AOTU(R)"           "BU(R)"            
#>  [70] "MB(+ACA)(R)"       "AL-DM2(R)"         "ATL(L)"           
#>  [73] "GA(R)"             "MB(L)"             "PB(L6)"           
#>  [76] "EBr3pw"            "b1(R)"             "AL-VC4(R)"        
#>  [79] "b'2(R)"            "AL-DM4(L)"         "ICL(L)"           
#>  [82] "ICL(R)"            "AL-DC1(R)"         "AL-DM5(R)"        
#>  [85] "AL-DM6(R)"         "AMMC"              "g4(R)"            
#>  [88] "NO(R)"             "CRE(L)"            "AB(L)"            
#>  [91] "CRE(-RUB)(L)"      "NO2(L)"            "AL-DA3(L)"        
#>  [94] "AL-DL5(L)"         "FBl6"              "AL-VL2p(R)"       
#>  [97] "NO1(R)"            "PB(L2)"            "LH(R)"            
#> [100] "GOR(R)"            "AL-VL1(R)"         "vACA(R)"          
#> [103] "AL-DL4(L)"         "AL-DA3(R)"         "CRE(-ROB,-RUB)(R)"
#> [106] "PB(R6)"            "AB(R)"             "AL-VM4(R)"        
#> [109] "FLA(R)"            "PB(L7)"            "LO(R)"            
#> [112] "AL-DL3(R)"         "MB(R)"             "AL-DP1l(R)"       
#> [115] "AL-VP1d(R)"        "mALT(R)"           "WED(R)"           
#> [118] "PB(L1)"            "VLNP(R)"           "AL-VP5(R)"        
#> [121] "PRW"               "b2(R)"             "EBr5"             
#> [124] "a'L(R)"            "NO3(R)"            "AL-DM1(L)"        
#> [127] "AL-VA6(R)"         "CRE(R)"            "AL-DM5(L)"        
#> [130] "a1(R)"             "AL-VP4(R)"         "SPS(L)"           
#> [133] "a'2(R)"            "ROB(R)"            "AL-DL1(R)"        
#> [136] "NO2(R)"            "AL-VP1l(R)"        "LOP(R)"           
#> [139] "GNG"               "OL(R)"             "AL-VC3l(R)"       
#> [142] "g3(R)"             "EBr6"              "AL-DP1m(L)"       
#> [145] "PB(L9)"            "SCL(R)"            "AL-VP1m(R)"       
#> [148] "SMP(R)"            "PENP"              "ME(R)"            
#> [151] "AL(L)"             "AL-DA4l(R)"        "AL-VA7m(R)"       
#> [154] "AL-DM6(L)"         "PB(R3)"            "AL-VC1(R)"        
#> [157] "AL-VM3(R)"         "mALT(L)"           "PB(R8)"           
#> [160] "FBl2"              "FBl8"              "CA(R)"            
#> [163] "AL-VM7v(R)"        "GOR(L)"            "FBl9"             
#> [166] "PB(L4)"            "AL-VC3m(R)"        "AL-VM7v(L)"       
#> [169] "CX"                "a2(R)"             "PB(L8)"           
#> [172] "AL-VA6(L)"         "EBr3d"             "AL-DC4(R)"        
#> [175] "AL-DL4(R)"         "FBl5"              "AL-DA1(R)"        
#> [178] "AL-DL2v(R)"        "gL(L)"             "AL-DC4(L)"        
#> [181] "SNP(L)"            "AL-VP2(R)"         "IPS(R)"           
#> [184] "FB"                "SNP(R)"            "AL-DA4m(R)"       
#> [187] "AL-VC2(R)"         "PED(R)"            "AL-DA4m(L)"       
#> [190] "AL-VM7d(R)"        "AOT(R)"            "aL(L)"            
#> [193] "AL-DC2(L)"         "CAN(R)"            "b'L(L)"           
#> [196] "AL-VA3(R)"         "PLP(R)"            "a3(R)"            
#> [199] "AL(R)"             "EBr1"              "SIP(R)"           
#> [202] "VES(R)"            "ATL(R)"            "dACA(R)"          
#> [205] "AME(R)"            "bL(R)"             "SCL(L)"           
#> [208] "PB(L3)"            "AL-VM5d(R)"        "b'1(R)"           
#> [211] "IB"                "PB(R7)"            "PVLP(R)"          
#> [214] "AL-DM1(R)"         "EBr3am"            "EBr2r4"           
#> [217] "g5(R)"             "AL-VA5(R)"         "FB-column3"       
#> [220] "LAL(R)"            "AL-VM5v(R)"        "AL-DM3(R)"        
#> [223] "AL-D(R)"           "AL-VL2a(R)"        "SIP(L)"           
#> [226] "AL-DC1(L)"         "FBl7"              "AL-DL5(R)"        
#> [229] "NO3(L)"            "CA(L)"            
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "SNP(R)"            "VLNP(R)"           "IB"               
#>   [4] "EBr2r4"            "ICL(R)"            "LX(L)"            
#>   [7] "b'L(L)"            "b'1(R)"            "AL-DM3(L)"        
#>  [10] "FB-column3"        "AOT(R)"            "SMP(R)"           
#>  [13] "a1(R)"             "a2(R)"             "mALT(R)"          
#>  [16] "BU(R)"             "PB(R2)"            "AL-DM4(R)"        
#>  [19] "AL-VC3l(R)"        "bL(R)"             "g4(R)"            
#>  [22] "AL(L)"             "AL-DM6(L)"         "b1(R)"            
#>  [25] "AL-DA3(L)"         "AL-D(L)"           "PB(L5)"           
#>  [28] "AL-VL1(R)"         "PENP"              "AL-DA2(L)"        
#>  [31] "PB(R9)"            "ICL(L)"            "VES(R)"           
#>  [34] "GNG"               "AL-DC4(R)"         "AL-VM5v(R)"       
#>  [37] "GF(R)"             "g5(R)"             "PB(L4)"           
#>  [40] "OL(R)"             "AL-VC5(R)"         "NO2(L)"           
#>  [43] "SMP(L)"            "AL-DM1(R)"         "VMNP"             
#>  [46] "AL-VP5(R)"         "g3(R)"             "EB"               
#>  [49] "AL-DM3(R)"         "AL-VM7d(R)"        "CA(L)"            
#>  [52] "b2(R)"             "AL-DC2(R)"         "AL-VM7v(R)"       
#>  [55] "POC"               "PB(L1)"            "PB(L8)"           
#>  [58] "AL-DA4m(L)"        "PVLP(R)"           "LH(R)"            
#>  [61] "RUB(L)"            "AL-DP1m(R)"        "AL-DC4(L)"        
#>  [64] "AL-DM4(L)"         "PB(L9)"            "FBl9"             
#>  [67] "AL-VA1d(R)"        "b'2(R)"            "LAL(-GA)(R)"      
#>  [70] "AL-DC1(L)"         "AL-VP1m(R)"        "AB(L)"            
#>  [73] "LX(R)"             "AL-VA6(L)"         "PB(R1)"           
#>  [76] "PB(R7)"            "lACA(R)"           "FBl8"             
#>  [79] "SAD"               "AL-VM7v(L)"        "NO"               
#>  [82] "AL-DL4(R)"         "FBl2"              "AL-DA4l(R)"       
#>  [85] "AL-VP1l(R)"        "BU(L)"             "SPS(L)"           
#>  [88] "LAL(L)"            "g1(R)"             "EBr3d"            
#>  [91] "ATL(L)"            "SIP(L)"            "AL-DM5(L)"        
#>  [94] "AL-DL5(R)"         "CRE(R)"            "NO1(R)"           
#>  [97] "AL-DM6(R)"         "RUB(R)"            "SCL(L)"           
#> [100] "AL-VL2p(R)"        "AL-VM1(R)"         "a'2(R)"           
#> [103] "FBl4"              "a'1(R)"            "PB(R3)"           
#> [106] "LAL(R)"            "AL-DL2v(R)"        "PB(R5)"           
#> [109] "WED(R)"            "MB(L)"             "dACA(R)"          
#> [112] "FLA(R)"            "AL-DL3(R)"         "AL-VA3(R)"        
#> [115] "vACA(R)"           "aL(R)"             "a'L(R)"           
#> [118] "NO(R)"             "gL(L)"             "AL-VC4(R)"        
#> [121] "AL-V(R)"           "LOP(R)"            "AMMC"             
#> [124] "NO(L)"             "AL-VA4(R)"         "AL-VA6(R)"        
#> [127] "AL-VP4(R)"         "GC"                "AL-DP1m(L)"       
#> [130] "a'L(L)"            "MB(R)"             "AL-DL2d(R)"       
#> [133] "CRE(-ROB,-RUB)(R)" "PB(R6)"            "PRW"              
#> [136] "AL-DA4m(R)"        "mALT(L)"           "AL-DL5(L)"        
#> [139] "AL-VA7m(R)"        "PB(L3)"            "NO1(L)"           
#> [142] "AB(R)"             "AME(R)"            "VES(L)"           
#> [145] "AL-VM2(R)"         "SPS(R)"            "g2(R)"            
#> [148] "EPA(R)"            "AL-VM4(R)"         "AL-VP1d(R)"       
#> [151] "AL-DC1(R)"         "EBr6"              "EBr5"             
#> [154] "PB(R8)"            "PB(L6)"            "bL(L)"            
#> [157] "NO3(L)"            "ATL(R)"            "PB(L7)"           
#> [160] "AL-VA1v(R)"        "AL-DL4(L)"         "SNP(L)"           
#> [163] "AL-DA1(R)"         "PB"                "AL-DC2(L)"        
#> [166] "LO(R)"             "AL-DL1(R)"         "AL-VA7l(R)"       
#> [169] "CA(R)"             "FB"                "AL-DM1(L)"        
#> [172] "IPS(R)"            "AL-VL2a(R)"        "PED(R)"           
#> [175] "SIP(R)"            "FBl1"              "a'3(R)"           
#> [178] "hemibrain"         "INP"               "AOTU(R)"          
#> [181] "AL-DA3(R)"         "AL-D(R)"           "AL-VA2(R)"        
#> [184] "AL-VP2(R)"         "CX"                "CRE(L)"           
#> [187] "SLP(R)"            "ME(R)"             "CRE(-RUB)(L)"     
#> [190] "MB(+ACA)(R)"       "NO2(R)"            "PLP(R)"           
#> [193] "AVLP(R)"           "PB(R4)"            "AL-VC1(R)"        
#> [196] "EBr3pw"            "NO3(R)"            "ROB(R)"           
#> [199] "FBl3"              "AL-VC3m(R)"        "b'L(R)"           
#> [202] "AL-DM2(L)"         "aL(L)"             "a3(R)"            
#> [205] "AL-DP1l(R)"        "AL-VP3(R)"         "AL-VM7d(L)"       
#> [208] "FBl7"              "AL-DC3(R)"         "AL-DM2(R)"        
#> [211] "AL-VC2(R)"         "PB(L2)"            "GA(R)"            
#> [214] "AL(R)"             "SAD(-AMMC)"        "CAN(R)"           
#> [217] "EBr1"              "gL(R)"             "EPA(L)"           
#> [220] "SCL(R)"            "AL-DM5(R)"         "AL-VM3(R)"        
#> [223] "GOR(L)"            "FBl5"              "GOR(R)"           
#> [226] "AL-DA2(R)"         "AL-VM5d(R)"        "FBl6"             
#> [229] "AL-VA5(R)"         "EBr3am"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method   path                                      
#> [1,] "DELETE" "/api"                                    
#> [2,] "PATCH"  "/api/*"                                  
#> [3,] "GET"    "/api/npexplorer/autapses"                
#> [4,] "GET"    "/token"                                  
#> [5,] "POST"   "/api/npexplorer/neuronmeta"              
#> [6,] "POST"   "/api/v:ver/npexplorer/commonconnectivity"
#>      name                                                                                      
#> [1,] "github.com/labstack/echo/v4.glob..func1"                                                 
#> [2,] "github.com/labstack/echo/v4.glob..func1"                                                 
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"  
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                  
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMeta-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                      
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }