Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DM4(L)"         "MB(+ACA)(R)"       "b1(R)"            
#>   [4] "AL-DP1m(R)"        "a'1(R)"            "FB-column3"       
#>   [7] "AL-DC4(R)"         "LOP(R)"            "AL-DL4(R)"        
#>  [10] "BU(R)"             "PB(L7)"            "AL-DL5(R)"        
#>  [13] "AL-VA4(R)"         "ROB(R)"            "CRE(-RUB)(L)"     
#>  [16] "FBl7"              "VLNP(R)"           "SPS(L)"           
#>  [19] "SNP(L)"            "AL-VM4(R)"         "AL-VM7d(R)"       
#>  [22] "gL(R)"             "b'L(L)"            "b'1(R)"           
#>  [25] "dACA(R)"           "AL-D(R)"           "AL-VC3l(R)"       
#>  [28] "GF(R)"             "LAL(-GA)(R)"       "AL-DM1(R)"        
#>  [31] "AL-VA6(R)"         "g1(R)"             "PVLP(R)"          
#>  [34] "PB(R2)"            "a1(R)"             "vACA(R)"          
#>  [37] "WED(R)"            "EBr3d"             "AL-VM1(R)"        
#>  [40] "AL-VM3(R)"         "AL-VM5d(R)"        "EBr3am"           
#>  [43] "AL-D(L)"           "FBl9"              "IPS(R)"           
#>  [46] "AME(R)"            "a'L(L)"            "hemibrain"        
#>  [49] "EBr5"              "AL-DL4(L)"         "ICL(R)"           
#>  [52] "SCL(R)"            "NO(L)"             "AL-VM2(R)"        
#>  [55] "CA(L)"             "b2(R)"             "g5(R)"            
#>  [58] "IB112"             "RUB(L)"            "AL(R)"            
#>  [61] "PB(L2)"            "SLP(R)"            "AL-DM2(R)"        
#>  [64] "AL-DM3(R)"         "AL-DM4(R)"         "MB(L)"            
#>  [67] "PB(L8)"            "AL-VM7v(L)"        "FBl4"             
#>  [70] "AL-DM5(L)"         "NO3(L)"            "EBr1"             
#>  [73] "EBr6"              "PLP(R)"            "a'L(R)"           
#>  [76] "NO(R)"             "AL-DC1(R)"         "ATL(L)"           
#>  [79] "a2(R)"             "AL-DC4(L)"         "aL(L)"            
#>  [82] "AL-VC4(R)"         "FB"                "AL-VA5(R)"        
#>  [85] "b'L(R)"            "EBr2r4"            "EPA(L)"           
#>  [88] "INP"               "AL-DA2(R)"         "AL-DA4m(R)"       
#>  [91] "AL-VA1v(R)"        "AL-VA2(R)"         "SIP(R)"           
#>  [94] "PB(L9)"            "FBl2"              "AB(R)"            
#>  [97] "AB(L)"             "a3(R)"             "PB(L4)"           
#> [100] "AL-DC3(R)"         "AL-DL2v(R)"        "bL(R)"            
#> [103] "AL-DA3(R)"         "AL-VA3(R)"         "b'2(R)"           
#> [106] "gL(L)"             "AL-DC2(L)"         "NO3(R)"           
#> [109] "a'2(R)"            "PB(L5)"            "NO2(L)"           
#> [112] "VMNP"              "NO"                "AL-DL2d(R)"       
#> [115] "AL-DM1(L)"         "AL-VC1(R)"         "AL-VP4(R)"        
#> [118] "AL-DA1(R)"         "AL-VA1d(R)"        "AL-DA3(L)"        
#> [121] "FBl1"              "AL-DL1(R)"         "SCL(L)"           
#> [124] "CA(R)"             "LX(L)"             "AL-VP1d(R)"       
#> [127] "GOR(L)"            "FLA(R)"            "AL-DL5(L)"        
#> [130] "AL-VC3m(R)"        "PRW"               "SNP(R)"           
#> [133] "AL-VC2(R)"         "PB(L1)"            "AL-VA6(L)"        
#> [136] "PB(R9)"            "LAL(R)"            "RUB(R)"           
#> [139] "PB(R3)"            "AL-VM7v(R)"        "AL-V(R)"          
#> [142] "AL-VP1m(R)"        "PED(R)"            "PB(R8)"           
#> [145] "AL-DC1(L)"         "CRE(L)"            "PB"               
#> [148] "ICL(L)"            "OL(R)"             "MB(R)"            
#> [151] "BU(L)"             "AMMC"              "LO(R)"            
#> [154] "AL-DM6(R)"         "PB(R6)"            "AL-DL3(R)"        
#> [157] "LAL(L)"            "mALT(R)"           "AL-VL2p(R)"       
#> [160] "aL(R)"             "VES(R)"            "AL-DA4l(R)"       
#> [163] "AL-VC5(R)"         "AVLP(R)"           "CX"               
#> [166] "PENP"              "NO1(L)"            "NO1(R)"           
#> [169] "g4(R)"             "GC"                "PB(L6)"           
#> [172] "AL-VA7m(R)"        "VES(L)"            "AL-VL2a(R)"       
#> [175] "AL-DA4m(L)"        "AL-VM7d(L)"        "AL-VA7l(R)"       
#> [178] "AOT(R)"            "a'3(R)"            "SMP(L)"           
#> [181] "FBl6"              "CAN(R)"            "GNG"              
#> [184] "AL-DM5(R)"         "lACA(R)"           "EBr3pw"           
#> [187] "FBl5"              "AL-VP5(R)"         "CRE(R)"           
#> [190] "SAD(-AMMC)"        "FBl3"              "ATL(R)"           
#> [193] "NO2(R)"            "g3(R)"             "PB(L3)"           
#> [196] "SIP(L)"            "SPS(R)"            "CRE(-ROB,-RUB)(R)"
#> [199] "EB"                "PB(R1)"            "bL(L)"            
#> [202] "SMP(R)"            "AL-VL1(R)"         "SAD"              
#> [205] "AL-DP1m(L)"        "PB(R4)"            "GOR(R)"           
#> [208] "AL-DC2(R)"         "mALT(L)"           "LX(R)"            
#> [211] "ME(R)"             "AL(L)"             "AL-DM2(L)"        
#> [214] "AL-VP1l(R)"        "AL-DA2(L)"         "AL-DM6(L)"        
#> [217] "AL-VP3(R)"         "AL-VP2(R)"         "POC"              
#> [220] "AOTU(R)"           "GA(R)"             "PB(R7)"           
#> [223] "AL-DM3(L)"         "EPA(R)"            "LH(R)"            
#> [226] "AL-DP1l(R)"        "PB(R5)"            "AL-VM5v(R)"       
#> [229] "IB"                "g2(R)"             "FBl8"             
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(R5)"            "mALT(R)"           "NO"               
#>   [4] "FBl6"              "NO3(L)"            "PB(L7)"           
#>   [7] "SNP(L)"            "NO2(R)"            "GNG"              
#>  [10] "NO1(R)"            "SIP(R)"            "ROB(R)"           
#>  [13] "a3(R)"             "AB(R)"             "ATL(L)"           
#>  [16] "a1(R)"             "PB(R7)"            "VES(R)"           
#>  [19] "FBl8"              "SMP(R)"            "PB(L3)"           
#>  [22] "GOR(L)"            "RUB(L)"            "EB"               
#>  [25] "POC"               "LOP(R)"            "a'3(R)"           
#>  [28] "hemibrain"         "GC"                "SPS(L)"           
#>  [31] "EPA(L)"            "NO(L)"             "SLP(R)"           
#>  [34] "g2(R)"             "FLA(R)"            "a'2(R)"           
#>  [37] "LX(L)"             "PB(L4)"            "OL(R)"            
#>  [40] "AL(R)"             "NO2(L)"            "SAD"              
#>  [43] "gL(R)"             "FBl1"              "g5(R)"            
#>  [46] "ATL(R)"            "GOR(R)"            "SMP(L)"           
#>  [49] "LAL(-GA)(R)"       "VMNP"              "PB(R1)"           
#>  [52] "a2(R)"             "g1(R)"             "dACA(R)"          
#>  [55] "MB(L)"             "gL(L)"             "FBl9"             
#>  [58] "b1(R)"             "CAN(R)"            "g4(R)"            
#>  [61] "AL(L)"             "EBr3pw"            "b'L(L)"           
#>  [64] "CX"                "lACA(R)"           "PB"               
#>  [67] "AME(R)"            "bL(L)"             "FB"               
#>  [70] "EBr6"              "PB(L1)"            "PENP"             
#>  [73] "SAD(-AMMC)"        "FBl3"              "WED(R)"           
#>  [76] "vACA(R)"           "AMMC"              "bL(R)"            
#>  [79] "RUB(R)"            "PB(L5)"            "SIP(L)"           
#>  [82] "aL(R)"             "LO(R)"             "AVLP(R)"          
#>  [85] "ICL(R)"            "a'1(R)"            "CA(R)"            
#>  [88] "LH(R)"             "EBr2r4"            "EPA(R)"           
#>  [91] "PVLP(R)"           "VLNP(R)"           "CRE(-ROB,-RUB)(R)"
#>  [94] "NO3(R)"            "b2(R)"             "PB(R9)"           
#>  [97] "GA(R)"             "PED(R)"            "ICL(L)"           
#> [100] "PB(R3)"            "b'1(R)"            "SCL(L)"           
#> [103] "PB(L2)"            "b'L(R)"            "EBr3d"            
#> [106] "BU(R)"             "a'L(R)"            "EBr3am"           
#> [109] "b'2(R)"            "SPS(R)"            "PB(R8)"           
#> [112] "CA(L)"             "ME(R)"             "PB(L6)"           
#> [115] "IPS(R)"            "AB(L)"             "AOTU(R)"          
#> [118] "PB(L9)"            "PB(R4)"            "PLP(R)"           
#> [121] "NO(R)"             "FBl4"              "aL(L)"            
#> [124] "INP"               "PB(R6)"            "LAL(L)"           
#> [127] "PRW"               "AOT(R)"            "IB"               
#> [130] "mALT(L)"           "LX(R)"             "FBl2"             
#> [133] "PB(R2)"            "LAL(R)"            "VES(L)"           
#> [136] "NO1(L)"            "PB(L8)"            "MB(+ACA)(R)"      
#> [139] "BU(L)"             "FBl7"              "a'L(L)"           
#> [142] "SCL(R)"            "EBr1"              "SNP(R)"           
#> [145] "EBr5"              "CRE(L)"            "FBl5"             
#> [148] "MB(R)"             "GF(R)"             "g3(R)"            
#> [151] "CRE(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DM3(L)"         "AL-VA2(R)"         "AL-VC5(R)"        
#>   [4] "b2(R)"             "g4(R)"             "PB(R4)"           
#>   [7] "AL-DL1(R)"         "SIP(L)"            "AL-DM2(L)"        
#>  [10] "AL-DP1m(L)"        "AL-VM2(R)"         "AL-DA4m(L)"       
#>  [13] "AL-DL4(L)"         "INP"               "SLP(R)"           
#>  [16] "AL-VP1d(R)"        "mALT(L)"           "NO3(R)"           
#>  [19] "GOR(L)"            "g1(R)"             "AL-DL5(R)"        
#>  [22] "PENP"              "PVLP(R)"           "bL(R)"            
#>  [25] "AL-VM3(R)"         "SMP(L)"            "SNP(R)"           
#>  [28] "EPA(R)"            "AL-DC2(R)"         "AOTU(R)"          
#>  [31] "EBr5"              "AL-DA2(L)"         "g3(R)"            
#>  [34] "GOR(R)"            "SCL(L)"            "CA(L)"            
#>  [37] "AL-VM7d(L)"        "PB(R3)"            "EPA(L)"           
#>  [40] "gL(R)"             "AL(L)"             "NO1(L)"           
#>  [43] "SAD"               "EB"                "AL-VC2(R)"        
#>  [46] "a1(R)"             "GA(R)"             "AL-VL2p(R)"       
#>  [49] "LOP(R)"            "PB(L5)"            "a'L(L)"           
#>  [52] "ICL(R)"            "AL-DA4m(R)"        "AL-VC3m(R)"       
#>  [55] "AL-VP5(R)"         "gL(L)"             "AL-VP4(R)"        
#>  [58] "PLP(R)"            "AL-DM6(L)"         "VES(L)"           
#>  [61] "MB(L)"             "AL-VA3(R)"         "AL-DL2v(R)"       
#>  [64] "AL-DM3(R)"         "AL-DC2(L)"         "FBl2"             
#>  [67] "AL-VM5d(R)"        "PB(L2)"            "FB-column3"       
#>  [70] "CRE(-RUB)(L)"      "FBl1"              "AL-D(L)"          
#>  [73] "AB(R)"             "aL(R)"             "SNP(L)"           
#>  [76] "AL-DP1l(R)"        "AL-VA5(R)"         "BU(R)"            
#>  [79] "EBr1"              "bL(L)"             "a3(R)"            
#>  [82] "AL-DM2(R)"         "AL-VC3l(R)"        "b'2(R)"           
#>  [85] "a'L(R)"            "FBl3"              "FBl7"             
#>  [88] "AME(R)"            "AL-VA7l(R)"        "AL-VL1(R)"        
#>  [91] "AL-VM1(R)"         "LAL(L)"            "FBl5"             
#>  [94] "FBl6"              "IPS(R)"            "GF(R)"            
#>  [97] "ATL(R)"            "AL-DC4(R)"         "CX"               
#> [100] "PRW"               "AL-VC1(R)"         "AB(L)"            
#> [103] "AL-VA7m(R)"        "dACA(R)"           "AL-VL2a(R)"       
#> [106] "EBr6"              "LX(L)"             "RUB(L)"           
#> [109] "NO(L)"             "AL-DA4l(R)"        "AL-VM5v(R)"       
#> [112] "AL-VM7v(R)"        "BU(L)"             "NO2(L)"           
#> [115] "AL-VP2(R)"         "b'L(R)"            "g2(R)"            
#> [118] "NO1(R)"            "AL-VM7v(L)"        "NO(R)"            
#> [121] "AL(R)"             "AL-VC4(R)"         "FB"               
#> [124] "PED(R)"            "FBl4"              "AL-D(R)"          
#> [127] "AL-VP3(R)"         "PB(R7)"            "EBr3pw"           
#> [130] "AL-DC1(R)"         "ATL(L)"            "AL-DL5(L)"        
#> [133] "aL(L)"             "AVLP(R)"           "VES(R)"           
#> [136] "PB(L4)"            "AL-VA1d(R)"        "EBr2r4"           
#> [139] "mALT(R)"           "PB(L1)"            "AL-DC4(L)"        
#> [142] "CA(R)"             "b'L(L)"            "AL-VP1l(R)"       
#> [145] "SMP(R)"            "PB(L3)"            "AL-DL4(R)"        
#> [148] "AL-DA2(R)"         "FLA(R)"            "ME(R)"            
#> [151] "CRE(L)"            "FBl8"              "AL-VA4(R)"        
#> [154] "SCL(R)"            "a'1(R)"            "ICL(L)"           
#> [157] "FBl9"              "AOT(R)"            "IB"               
#> [160] "PB"                "SAD(-AMMC)"        "PB(R1)"           
#> [163] "AL-VA1v(R)"        "PB(L6)"            "AL-DL2d(R)"       
#> [166] "AL-VM7d(R)"        "POC"               "WED(R)"           
#> [169] "AL-DA3(R)"         "MB(+ACA)(R)"       "PB(R8)"           
#> [172] "AL-DM5(L)"         "NO3(L)"            "LO(R)"            
#> [175] "RUB(R)"            "PB(L8)"            "LAL(-GA)(R)"      
#> [178] "g5(R)"             "AL-DM1(R)"         "hemibrain"        
#> [181] "b1(R)"             "AL-DC1(L)"         "AL-DM1(L)"        
#> [184] "a'2(R)"            "PB(R9)"            "ROB(R)"           
#> [187] "vACA(R)"           "CRE(-ROB,-RUB)(R)" "LX(R)"            
#> [190] "AL-VP1m(R)"        "AMMC"              "AL-DA1(R)"        
#> [193] "AL-VA6(L)"         "PB(L9)"            "lACA(R)"          
#> [196] "AL-DP1m(R)"        "NO"                "a'3(R)"           
#> [199] "OL(R)"             "PB(R2)"            "GNG"              
#> [202] "AL-DC3(R)"         "PB(R6)"            "NO2(R)"           
#> [205] "AL-VA6(R)"         "EBr3am"            "PB(L7)"           
#> [208] "AL-DM5(R)"         "PB(R5)"            "AL-VM4(R)"        
#> [211] "GC"                "EBr3d"             "AL-DA3(L)"        
#> [214] "SIP(R)"            "AL-DL3(R)"         "SPS(R)"           
#> [217] "AL-V(R)"           "a2(R)"             "AL-DM4(L)"        
#> [220] "CAN(R)"            "AL-DM6(R)"         "VLNP(R)"          
#> [223] "VMNP"              "LH(R)"             "LAL(R)"           
#> [226] "AL-DM4(R)"         "b'1(R)"            "CRE(R)"           
#> [229] "MB(R)"             "SPS(L)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "hemibrain"         "AL-VC3l(R)"        "SCL(L)"           
#>   [4] "AL-DC2(L)"         "AL-D(R)"           "mALT(L)"          
#>   [7] "PB(L5)"            "g3(R)"             "gL(L)"            
#>  [10] "AL-DP1m(L)"        "b'1(R)"            "AL-VP2(R)"        
#>  [13] "a2(R)"             "EBr3d"             "SIP(R)"           
#>  [16] "AL-VC1(R)"         "AL-VP4(R)"         "a'1(R)"           
#>  [19] "AL-DL4(R)"         "AL-DM2(R)"         "AL-VM4(R)"        
#>  [22] "GC"                "FB"                "PB(R8)"           
#>  [25] "PB"                "SMP(L)"            "LAL(-GA)(R)"      
#>  [28] "LX(L)"             "AL-VM5d(R)"        "AL-DM3(L)"        
#>  [31] "AL-DA2(L)"         "AL-DM6(L)"         "LH(R)"            
#>  [34] "AL-VA3(R)"         "EBr3am"            "PLP(R)"           
#>  [37] "AL-DA4m(R)"        "AL-VM2(R)"         "AL-V(R)"          
#>  [40] "BU(L)"             "g4(R)"             "FBl9"             
#>  [43] "SAD"               "PB(R2)"            "AL-VM1(R)"        
#>  [46] "EB"                "GNG"               "AL-DL1(R)"        
#>  [49] "AL-DP1l(R)"        "AL-DM4(R)"         "AL-VA7l(R)"       
#>  [52] "AL-VL1(R)"         "AOTU(R)"           "FB-column3"       
#>  [55] "PB(R3)"            "INP"               "AL-VA6(L)"        
#>  [58] "AL-VA7m(R)"        "PB(L1)"            "bL(R)"            
#>  [61] "AL-VL2p(R)"        "AL-DC3(R)"         "AL-DP1m(R)"       
#>  [64] "AL-DM4(L)"         "AL-VM7d(L)"        "GF(R)"            
#>  [67] "EBr2r4"            "AL-VM3(R)"         "CAN(R)"           
#>  [70] "PB(L4)"            "AL-VL2a(R)"        "AB(L)"            
#>  [73] "PB(R7)"            "EBr6"              "AL-DL4(L)"        
#>  [76] "NO(R)"             "AMMC"              "PB(L3)"           
#>  [79] "AL-DM1(R)"         "AL-VA5(R)"         "AL-VM7d(R)"       
#>  [82] "AL-VA1d(R)"        "AL-VA6(R)"         "AOT(R)"           
#>  [85] "a'3(R)"            "SPS(R)"            "ROB(R)"           
#>  [88] "AL-DA2(R)"         "AL-DL5(R)"         "AL-DM5(R)"        
#>  [91] "VES(R)"            "AL-VP1d(R)"        "BU(R)"            
#>  [94] "ICL(L)"            "AL-VA4(R)"         "LAL(L)"           
#>  [97] "NO3(L)"            "NO(L)"             "CA(L)"            
#> [100] "a'L(R)"            "PB(L2)"            "CX"               
#> [103] "b1(R)"             "AB(R)"             "PB(R5)"           
#> [106] "GOR(L)"            "dACA(R)"           "FLA(R)"           
#> [109] "FBl7"              "GOR(R)"            "AL-DL2v(R)"       
#> [112] "AL-VP1l(R)"        "GA(R)"             "AL-DC1(L)"        
#> [115] "g5(R)"             "NO2(R)"            "VLNP(R)"          
#> [118] "PENP"              "g1(R)"             "EBr1"             
#> [121] "AL-DA4m(L)"        "MB(L)"             "LX(R)"            
#> [124] "PRW"               "FBl2"              "AVLP(R)"          
#> [127] "vACA(R)"           "EPA(R)"            "AL-VM7v(L)"       
#> [130] "PVLP(R)"           "FBl1"              "AL-DC4(L)"        
#> [133] "FBl4"              "ME(R)"             "AL-VA2(R)"        
#> [136] "CRE(-ROB,-RUB)(R)" "SNP(R)"            "AL-DM1(L)"        
#> [139] "CRE(L)"            "PB(L7)"            "PED(R)"           
#> [142] "IPS(R)"            "AL-VC2(R)"         "PB(R6)"           
#> [145] "SNP(L)"            "SLP(R)"            "AL-DC1(R)"        
#> [148] "AL-VC5(R)"         "NO1(R)"            "PB(R9)"           
#> [151] "FBl6"              "FBl8"              "aL(R)"            
#> [154] "RUB(L)"            "b'L(L)"            "AL-DM2(L)"        
#> [157] "FBl3"              "AL-DC2(R)"         "AME(R)"           
#> [160] "AL-DA4l(R)"        "AL-VP1m(R)"        "POC"              
#> [163] "RUB(R)"            "mALT(R)"           "CRE(R)"           
#> [166] "LO(R)"             "AL(R)"             "NO"               
#> [169] "PB(L6)"            "IB"                "AL-VP5(R)"        
#> [172] "LOP(R)"            "NO2(L)"            "b'2(R)"           
#> [175] "WED(R)"            "AL-DA1(R)"         "AL-VC4(R)"        
#> [178] "g2(R)"             "AL-DL5(L)"         "FBl5"             
#> [181] "AL-DL3(R)"         "AL-DM3(R)"         "lACA(R)"          
#> [184] "NO3(R)"            "PB(L8)"            "SAD(-AMMC)"       
#> [187] "AL-VC3m(R)"        "SPS(L)"            "gL(R)"            
#> [190] "OL(R)"             "AL-DM6(R)"         "SIP(L)"           
#> [193] "a3(R)"             "AL-DM5(L)"         "VES(L)"           
#> [196] "b2(R)"             "PB(L9)"            "CRE(-RUB)(L)"     
#> [199] "bL(L)"             "PB(R4)"            "LAL(R)"           
#> [202] "a'L(L)"            "a'2(R)"            "AL(L)"            
#> [205] "MB(+ACA)(R)"       "ATL(R)"            "AL-DA3(L)"        
#> [208] "NO1(L)"            "PB(R1)"            "ATL(L)"           
#> [211] "a1(R)"             "SMP(R)"            "VMNP"             
#> [214] "AL-D(L)"           "AL-DA3(R)"         "AL-VA1v(R)"       
#> [217] "AL-DC4(R)"         "AL-VM7v(R)"        "SCL(R)"           
#> [220] "EBr5"              "AL-VP3(R)"         "MB(R)"            
#> [223] "AL-VM5v(R)"        "EBr3pw"            "EPA(L)"           
#> [226] "AL-DL2d(R)"        "b'L(R)"            "CA(R)"            
#> [229] "ICL(R)"            "aL(L)"            
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method  path                                
#> [1,] "POST"  "/api/v:ver/npexplorer/distribution"
#> [2,] "POST"  "/api/npexplorer/completeness"      
#> [3,] "TRACE" "/api/*"                            
#> [4,] "POST"  "/api/npexplorer/roisinneuron"      
#> [5,] "GET"   "/api/npexplorer/commonconnectivity"
#> [6,] "POST"  "/api"                              
#>      name                                                                                              
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCompleteness-fm"      
#> [3,] "github.com/labstack/echo/v4.glob..func1"                                                         
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"      
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCommonConnectivity-fm"
#> [6,] "github.com/labstack/echo/v4.glob..func1"                                                         
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }