Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $fib19
#> $fib19$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $fib19$uuid
#> [1] "6c1b926af14f43cbbcf85da97c46fb2b"
#> 
#> $fib19$ROIs
#> [1] "LOP" "ME"  "LO" 
#> 
#> $fib19$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $fib19$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $fib19$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "PB(R6)"            "g4(R)"             "AL-DC3(R)"        
#>   [4] "SCL(R)"            "LAL(R)"            "AL-DL5(R)"        
#>   [7] "AL-VP3(R)"         "SCL(L)"            "PED(R)"           
#>  [10] "AL-DM1(L)"         "a3(R)"             "NO2(R)"           
#>  [13] "GOR(R)"            "AL-VC3m(R)"        "GA(R)"            
#>  [16] "SIP(L)"            "aL(R)"             "EBr3pw"           
#>  [19] "FBl6"              "AL-VL2a(R)"        "EBr3am"           
#>  [22] "g2(R)"             "ME(R)"             "PB(L9)"           
#>  [25] "SNP(L)"            "VLNP(R)"           "g1(R)"            
#>  [28] "AME(R)"            "a'L(L)"            "ATL(L)"           
#>  [31] "AL-DC2(L)"         "AL-DL4(L)"         "AL-VA1v(R)"       
#>  [34] "AL-VC1(R)"         "SAD"               "NO1(L)"           
#>  [37] "AL-VA5(R)"         "a1(R)"             "SPS(R)"           
#>  [40] "a'2(R)"            "SNP(R)"            "AL-VM3(R)"        
#>  [43] "b2(R)"             "PB(L8)"            "FBl1"             
#>  [46] "ICL(R)"            "AL-DM4(R)"         "NO1(R)"           
#>  [49] "EPA(R)"            "IB"                "AL-DM3(L)"        
#>  [52] "AL-VM7d(L)"        "AL-VM7v(L)"        "AB(R)"            
#>  [55] "AL-VM5v(R)"        "VES(R)"            "GNG"              
#>  [58] "CRE(-ROB,-RUB)(R)" "EBr5"              "gL(R)"            
#>  [61] "FBl3"              "FBl8"              "AL-VP2(R)"        
#>  [64] "EB"                "AL-VA1d(R)"        "AL-VP1d(R)"       
#>  [67] "LAL(-GA)(R)"       "WED(R)"            "PB(R1)"           
#>  [70] "b'L(L)"            "VES(L)"            "EBr1"             
#>  [73] "LX(L)"             "vACA(R)"           "a2(R)"            
#>  [76] "FBl2"              "PB(L5)"            "a'L(R)"           
#>  [79] "AL(L)"             "AL-DP1m(L)"        "AL-VL1(R)"        
#>  [82] "EPA(L)"            "AL-VC2(R)"         "AL-VP4(R)"        
#>  [85] "mALT(L)"           "AL-DC4(R)"         "b'L(R)"           
#>  [88] "AL-DA4m(L)"        "PB(R2)"            "ICL(L)"           
#>  [91] "b1(R)"             "g3(R)"             "dACA(R)"          
#>  [94] "CA(R)"             "PB(L2)"            "PB(R7)"           
#>  [97] "EBr3d"             "PB(L6)"            "bL(L)"            
#> [100] "AL-VA7m(R)"        "PB(R9)"            "AL-VC4(R)"        
#> [103] "GOR(L)"            "GC"                "FB-column3"       
#> [106] "AL-VM7v(R)"        "AL-DM1(R)"         "AB(L)"            
#> [109] "VMNP"              "NO"                "ATL(R)"           
#> [112] "AL(R)"             "AL-DA4l(R)"        "BU(L)"            
#> [115] "FBl4"              "FBl7"              "CAN(R)"           
#> [118] "FB"                "LO(R)"             "AL-VA6(R)"        
#> [121] "PB(L1)"            "AL-DM4(L)"         "MB(+ACA)(R)"      
#> [124] "NO(L)"             "SLP(R)"            "AL-VA4(R)"        
#> [127] "AL-DA3(L)"         "AL-DM6(L)"         "AL-VL2p(R)"       
#> [130] "AL-DM2(L)"         "IB112"             "RUB(L)"           
#> [133] "MB(L)"             "EBr6"              "PLP(R)"           
#> [136] "AL-DA2(L)"         "AL-DM6(R)"         "AL-VP5(R)"        
#> [139] "SMP(R)"            "hemibrain"         "PRW"              
#> [142] "EBr2r4"            "gL(L)"             "AL-VC3l(R)"       
#> [145] "SMP(L)"            "aL(L)"             "AL-DA1(R)"        
#> [148] "b'2(R)"            "PB(R5)"            "lACA(R)"          
#> [151] "PB(L7)"            "bL(R)"             "AL-DL2v(R)"       
#> [154] "AL-DL4(R)"         "a'1(R)"            "IPS(R)"           
#> [157] "AL-DM5(R)"         "AL-DP1l(R)"        "AL-VM1(R)"        
#> [160] "AL-D(R)"           "AL-DL1(R)"         "ROB(R)"           
#> [163] "AL-DM3(R)"         "a'3(R)"            "AL-VM5d(R)"       
#> [166] "AL-VM7d(R)"        "AMMC"              "CRE(R)"           
#> [169] "BU(R)"             "CRE(L)"            "PB(R3)"           
#> [172] "AL-DA3(R)"         "RUB(R)"            "LOP(R)"           
#> [175] "CA(L)"             "NO2(L)"            "AOT(R)"           
#> [178] "b'1(R)"            "LAL(L)"            "NO3(R)"           
#> [181] "AL-D(L)"           "PB(R4)"            "POC"              
#> [184] "PB(L3)"            "AL-DC4(L)"         "FBl9"             
#> [187] "AL-VP1m(R)"        "AOTU(R)"           "LX(R)"            
#> [190] "NO3(L)"            "AL-DL3(R)"         "AL-DP1m(R)"       
#> [193] "PENP"              "SAD(-AMMC)"        "AL-DA4m(R)"       
#> [196] "AL-DM5(L)"         "CRE(-RUB)(L)"      "INP"              
#> [199] "PB(L4)"            "AL-VM2(R)"         "AL-DC1(L)"        
#> [202] "AL-DL5(L)"         "AL-VA3(R)"         "LH(R)"            
#> [205] "AL-VA7l(R)"        "AL-VP1l(R)"        "PB(R8)"           
#> [208] "SIP(R)"            "PVLP(R)"           "AL-DM2(R)"        
#> [211] "AL-VC5(R)"         "AL-VM4(R)"         "AL-VA6(L)"        
#> [214] "AL-DC1(R)"         "AL-V(R)"           "PB"               
#> [217] "FBl5"              "AL-DC2(R)"         "AL-DL2d(R)"       
#> [220] "mALT(R)"           "OL(R)"             "MB(R)"            
#> [223] "SPS(L)"            "NO(R)"             "g5(R)"            
#> [226] "AVLP(R)"           "AL-VA2(R)"         "FLA(R)"           
#> [229] "AL-DA2(R)"         "CX"                "GF(R)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "SMP(L)"            "gL(R)"             "ICL(R)"           
#>   [4] "PB(R7)"            "aL(R)"             "NO1(L)"           
#>   [7] "a'2(R)"            "AL(L)"             "GNG"              
#>  [10] "b'L(R)"            "g4(R)"             "VES(R)"           
#>  [13] "WED(R)"            "g2(R)"             "EBr5"             
#>  [16] "SPS(R)"            "bL(L)"             "VMNP"             
#>  [19] "EBr3d"             "SNP(L)"            "aL(L)"            
#>  [22] "AMMC"              "GOR(R)"            "a'L(L)"           
#>  [25] "PB(L5)"            "SAD"               "SIP(L)"           
#>  [28] "NO3(R)"            "FB"                "PB"               
#>  [31] "mALT(R)"           "ICL(L)"            "a3(R)"            
#>  [34] "LAL(R)"            "VLNP(R)"           "RUB(R)"           
#>  [37] "NO2(L)"            "AOTU(R)"           "MB(L)"            
#>  [40] "PB(L7)"            "LO(R)"             "b'1(R)"           
#>  [43] "SCL(L)"            "PB(R9)"            "PVLP(R)"          
#>  [46] "MB(+ACA)(R)"       "PB(R4)"            "OL(R)"            
#>  [49] "ATL(L)"            "PENP"              "EBr1"             
#>  [52] "SIP(R)"            "g5(R)"             "FBl1"             
#>  [55] "FBl3"              "AVLP(R)"           "SCL(R)"           
#>  [58] "BU(L)"             "FBl5"              "FBl7"             
#>  [61] "PB(R5)"            "gL(L)"             "BU(R)"            
#>  [64] "PB(L9)"            "FBl2"              "AB(R)"            
#>  [67] "RUB(L)"            "MB(R)"             "LOP(R)"           
#>  [70] "CRE(R)"            "NO1(R)"            "FBl6"             
#>  [73] "CAN(R)"            "EBr3am"            "GOR(L)"           
#>  [76] "dACA(R)"           "lACA(R)"           "GC"               
#>  [79] "PB(R3)"            "LX(L)"             "mALT(L)"          
#>  [82] "PLP(R)"            "PRW"               "b2(R)"            
#>  [85] "EB"                "INP"               "NO(L)"            
#>  [88] "SLP(R)"            "LAL(-GA)(R)"       "EBr6"             
#>  [91] "POC"               "PB(R8)"            "CRE(L)"           
#>  [94] "EPA(L)"            "IB"                "g3(R)"            
#>  [97] "PB(L8)"            "PB(L4)"            "AB(L)"            
#> [100] "vACA(R)"           "CA(L)"             "LAL(L)"           
#> [103] "a2(R)"             "hemibrain"         "PED(R)"           
#> [106] "IPS(R)"            "a'3(R)"            "PB(R6)"           
#> [109] "PB(L3)"            "GA(R)"             "g1(R)"            
#> [112] "SNP(R)"            "bL(R)"             "PB(R1)"           
#> [115] "SAD(-AMMC)"        "ATL(R)"            "b'L(L)"           
#> [118] "AOT(R)"            "SMP(R)"            "a1(R)"            
#> [121] "ME(R)"             "CA(R)"             "PB(L2)"           
#> [124] "CX"                "b'2(R)"            "LX(R)"            
#> [127] "PB(L6)"            "FBl4"              "AME(R)"           
#> [130] "EBr2r4"            "ROB(R)"            "FBl8"             
#> [133] "AL(R)"             "LH(R)"             "b1(R)"            
#> [136] "GF(R)"             "PB(L1)"            "FLA(R)"           
#> [139] "EPA(R)"            "NO2(R)"            "NO"               
#> [142] "a'1(R)"            "NO3(L)"            "EBr3pw"           
#> [145] "FBl9"              "PB(R2)"            "CRE(-ROB,-RUB)(R)"
#> [148] "SPS(L)"            "a'L(R)"            "NO(R)"            
#> [151] "VES(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "LX(R)"             "ME(R)"             "MB(L)"            
#>   [4] "FBl8"              "SLP(R)"            "gL(L)"            
#>   [7] "PB(L7)"            "PVLP(R)"           "AL-VC3l(R)"       
#>  [10] "EBr1"              "FBl5"              "aL(L)"            
#>  [13] "bL(R)"             "AL-DM5(R)"         "AL-VM5d(R)"       
#>  [16] "NO3(R)"            "CRE(-RUB)(L)"      "AL-D(L)"          
#>  [19] "FBl6"              "mALT(L)"           "PENP"             
#>  [22] "g1(R)"             "AL-V(R)"           "CX"               
#>  [25] "PLP(R)"            "AL-DM3(L)"         "AL-DA3(R)"        
#>  [28] "AL-VA1d(R)"        "AL-VA1v(R)"        "AL-VM7d(R)"       
#>  [31] "IB"                "PB(L3)"            "GOR(L)"           
#>  [34] "NO"                "b1(R)"             "hemibrain"        
#>  [37] "EPA(R)"            "PB(R9)"            "SMP(L)"           
#>  [40] "PRW"               "bL(L)"             "PB(R2)"           
#>  [43] "ATL(R)"            "AL-DM5(L)"         "GNG"              
#>  [46] "AL-VM4(R)"         "PB(L1)"            "AL-DC1(L)"        
#>  [49] "PB(R1)"            "FBl9"              "CA(R)"            
#>  [52] "SIP(L)"            "AL-DA4m(R)"        "GC"               
#>  [55] "FBl7"              "SPS(R)"            "AL-VC5(R)"        
#>  [58] "PB(R5)"            "AL-DA1(R)"         "RUB(R)"           
#>  [61] "a2(R)"             "SNP(L)"            "FB-column3"       
#>  [64] "AL-VM1(R)"         "NO2(L)"            "mALT(R)"          
#>  [67] "BU(L)"             "AME(R)"            "EPA(L)"           
#>  [70] "INP"               "AL-DA2(R)"         "AL-DA4l(R)"       
#>  [73] "AL-VP3(R)"         "b2(R)"             "VES(R)"           
#>  [76] "AL-DC1(R)"         "PB(R7)"            "FBl2"             
#>  [79] "SCL(L)"            "b'L(R)"            "ICL(L)"           
#>  [82] "GOR(R)"            "AL-DC4(R)"         "AL-VA5(R)"        
#>  [85] "AL-DA2(L)"         "AL-VC4(R)"         "a'L(L)"           
#>  [88] "b'L(L)"            "PB(L4)"            "MB(R)"            
#>  [91] "AL-DM2(R)"         "AL-DM4(R)"         "AL-VP1m(R)"       
#>  [94] "EBr3pw"            "LH(R)"             "AL-VP1d(R)"       
#>  [97] "AL-DL4(L)"         "AL-DM6(L)"         "gL(R)"            
#> [100] "AL-DL2v(R)"        "GF(R)"             "SPS(L)"           
#> [103] "AL-DL5(L)"         "g5(R)"             "a1(R)"            
#> [106] "EBr5"              "POC"               "FBl4"             
#> [109] "a3(R)"             "AL-DP1m(R)"        "AL-D(R)"          
#> [112] "AOT(R)"            "AL-VM7v(L)"        "NO1(L)"           
#> [115] "PB(R4)"            "AL-DM3(R)"         "PED(R)"           
#> [118] "OL(R)"             "AL-DM2(L)"         "NO1(R)"           
#> [121] "AL-DC2(L)"         "MB(+ACA)(R)"       "AL-VC3m(R)"       
#> [124] "PB(R6)"            "SIP(R)"            "NO(R)"            
#> [127] "AL-DL2d(R)"        "AMMC"              "g2(R)"            
#> [130] "IPS(R)"            "AL-VA3(R)"         "AL-VP5(R)"        
#> [133] "EBr3d"             "PB(R8)"            "SNP(R)"           
#> [136] "ATL(L)"            "CRE(R)"            "VES(L)"           
#> [139] "AL-DL3(R)"         "AL-DP1l(R)"        "AL-VA7l(R)"       
#> [142] "AL-VA7m(R)"        "AL-VM7d(L)"        "AL-DA3(L)"        
#> [145] "PB(L5)"            "AL-DC4(L)"         "AVLP(R)"          
#> [148] "NO(L)"             "b'1(R)"            "PB(L2)"           
#> [151] "AOTU(R)"           "a'L(R)"            "ROB(R)"           
#> [154] "PB(L6)"            "CRE(-ROB,-RUB)(R)" "FB"               
#> [157] "PB"                "AL(L)"             "AL-VA2(R)"        
#> [160] "AL-VA6(R)"         "AL-VP4(R)"         "AL-VP2(R)"        
#> [163] "AL-VP1l(R)"        "b'2(R)"            "CA(L)"            
#> [166] "FBl1"              "AL-DC2(R)"         "aL(R)"            
#> [169] "NO2(R)"            "AL-VC1(R)"         "AL-VM5v(R)"       
#> [172] "AL-VM7v(R)"        "LOP(R)"            "CRE(L)"           
#> [175] "ICL(R)"            "AL-VA6(L)"         "PB(L9)"           
#> [178] "a'1(R)"            "NO3(L)"            "SAD(-AMMC)"       
#> [181] "LAL(-GA)(R)"       "AL-DL5(R)"         "AL-DL1(R)"        
#> [184] "AL-DM1(L)"         "PB(R3)"            "RUB(L)"           
#> [187] "AL-VM3(R)"         "EBr6"              "AL-DA4m(L)"       
#> [190] "AL-VA4(R)"         "a'3(R)"            "VMNP"             
#> [193] "AL-DM1(R)"         "GA(R)"             "SMP(R)"           
#> [196] "AL-DC3(R)"         "g4(R)"             "AL-VL2p(R)"       
#> [199] "WED(R)"            "EB"                "LO(R)"            
#> [202] "AL-DL4(R)"         "dACA(R)"           "BU(R)"            
#> [205] "AL-DP1m(L)"        "EBr2r4"            "AL(R)"            
#> [208] "AL-DM6(R)"         "AL-VL2a(R)"        "EBr3am"           
#> [211] "PB(L8)"            "AB(L)"             "LAL(R)"           
#> [214] "FLA(R)"            "g3(R)"             "SCL(R)"           
#> [217] "FBl3"              "LX(L)"             "CAN(R)"           
#> [220] "LAL(L)"            "VLNP(R)"           "AL-VL1(R)"        
#> [223] "AB(R)"             "AL-VM2(R)"         "vACA(R)"          
#> [226] "lACA(R)"           "a'2(R)"            "AL-DM4(L)"        
#> [229] "AL-VC2(R)"         "SAD"              
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-VP5(R)"         "a'2(R)"            "AL-DL4(L)"        
#>   [4] "a'1(R)"            "NO(L)"             "AL-VC4(R)"        
#>   [7] "AL-DA4m(L)"        "FBl3"              "AL-DM3(R)"        
#>  [10] "AL-VA7m(R)"        "PB(L8)"            "SMP(L)"           
#>  [13] "CRE(L)"            "AL-VA1v(R)"        "AL-VA6(R)"        
#>  [16] "CRE(-ROB,-RUB)(R)" "hemibrain"         "AL-DC2(L)"        
#>  [19] "SIP(R)"            "PB(L9)"            "a3(R)"            
#>  [22] "BU(L)"             "FBl2"              "AL-DL5(R)"        
#>  [25] "AL-VM1(R)"         "a1(R)"             "LAL(L)"           
#>  [28] "MB(R)"             "NO1(R)"            "BU(R)"            
#>  [31] "b'1(R)"            "FBl1"              "g1(R)"            
#>  [34] "AL-DP1m(L)"        "NO(R)"             "gL(R)"            
#>  [37] "NO3(L)"            "SPS(R)"            "aL(L)"            
#>  [40] "VLNP(R)"           "LOP(R)"            "SCL(R)"           
#>  [43] "AL-DA3(L)"         "PB(R4)"            "AL-DP1m(R)"       
#>  [46] "AL-DM5(R)"         "GF(R)"             "VMNP"             
#>  [49] "AL-DM1(L)"         "AL-DA2(R)"         "AL-DM1(R)"        
#>  [52] "AL-VP4(R)"         "PLP(R)"            "EPA(R)"           
#>  [55] "PVLP(R)"           "AL-DA1(R)"         "AL-VM5v(R)"       
#>  [58] "AL-VA7l(R)"        "gL(L)"             "AL-DM4(L)"        
#>  [61] "FBl7"              "AL-DP1l(R)"        "AL-VC5(R)"        
#>  [64] "OL(R)"             "AL-DC3(R)"         "SPS(L)"           
#>  [67] "EBr2r4"            "AOT(R)"            "FBl6"             
#>  [70] "POC"               "AL-VC3l(R)"        "PRW"              
#>  [73] "PED(R)"            "AL-D(L)"           "AL-DM5(L)"        
#>  [76] "PB(L7)"            "PB(R6)"            "FB-column3"       
#>  [79] "EBr3d"             "PB(L3)"            "FBl9"             
#>  [82] "LH(R)"             "b1(R)"             "AL-DC2(R)"        
#>  [85] "SCL(L)"            "GC"                "IPS(R)"           
#>  [88] "AL-VC2(R)"         "AL-VL2a(R)"        "AL-VM3(R)"        
#>  [91] "mALT(R)"           "PB(L1)"            "GOR(R)"           
#>  [94] "b'2(R)"            "aL(R)"             "AL-VA2(R)"        
#>  [97] "LAL(-GA)(R)"       "MB(L)"             "AL-DL5(L)"        
#> [100] "ROB(R)"            "VES(R)"            "EBr6"             
#> [103] "AL-VC1(R)"         "FLA(R)"            "PB"               
#> [106] "EBr3pw"            "NO2(R)"            "AL-VM7v(R)"       
#> [109] "RUB(L)"            "ATL(L)"            "NO2(L)"           
#> [112] "AL-DC4(L)"         "SAD(-AMMC)"        "SMP(R)"           
#> [115] "AL-DM3(L)"         "AL-DM6(L)"         "b'L(L)"           
#> [118] "AL-VP1m(R)"        "AL-VP1l(R)"        "vACA(R)"          
#> [121] "PB(R7)"            "GOR(L)"            "AL-DC1(L)"        
#> [124] "CRE(-RUB)(L)"      "AL-DC1(R)"         "AL-VM7d(R)"       
#> [127] "LO(R)"             "ICL(L)"            "CA(R)"            
#> [130] "AL-DL1(R)"         "AL-VA1d(R)"        "AL-VP1d(R)"       
#> [133] "SAD"               "EB"                "ICL(R)"           
#> [136] "PB(L4)"            "AL-DL2d(R)"        "PENP"             
#> [139] "a'3(R)"            "PB(R8)"            "PB(R9)"           
#> [142] "ATL(R)"            "AL-VC3m(R)"        "b2(R)"            
#> [145] "AL-VA6(L)"         "SIP(L)"            "WED(R)"           
#> [148] "PB(R2)"            "AME(R)"            "AB(L)"            
#> [151] "mALT(L)"           "CRE(R)"            "SNP(R)"           
#> [154] "AL-VM7d(L)"        "PB(R1)"            "AVLP(R)"          
#> [157] "CA(L)"             "LX(R)"             "PB(L6)"           
#> [160] "NO"                "MB(+ACA)(R)"       "INP"              
#> [163] "VES(L)"            "PB(L2)"            "AL-DA4m(R)"       
#> [166] "AL-VA5(R)"         "AL-VP3(R)"         "IB"               
#> [169] "EPA(L)"            "GNG"               "AL-DL3(R)"        
#> [172] "AL-DL4(R)"         "AL-VP2(R)"         "NO3(R)"           
#> [175] "AMMC"              "lACA(R)"           "LX(L)"            
#> [178] "AL-VL1(R)"         "AL-VM5d(R)"        "NO1(L)"           
#> [181] "g5(R)"             "GA(R)"             "bL(L)"            
#> [184] "FBl5"              "PB(R3)"            "AB(R)"            
#> [187] "EBr1"              "EBr5"              "bL(R)"            
#> [190] "a2(R)"             "SNP(L)"            "RUB(R)"           
#> [193] "PB(L5)"            "b'L(R)"            "AL(L)"            
#> [196] "AL-DA2(L)"         "AL-DC4(R)"         "AL-D(R)"          
#> [199] "EBr3am"            "AL-VM7v(L)"        "SLP(R)"           
#> [202] "AL-DA4l(R)"        "AL-V(R)"           "g2(R)"            
#> [205] "AL-DM2(L)"         "PB(R5)"            "FB"               
#> [208] "a'L(R)"            "FBl4"              "CAN(R)"           
#> [211] "AL(R)"             "AL-DL2v(R)"        "AL-DM4(R)"        
#> [214] "g4(R)"             "AL-DM2(R)"         "AL-DM6(R)"        
#> [217] "AL-VM2(R)"         "AL-VM4(R)"         "g3(R)"            
#> [220] "FBl8"              "LAL(R)"            "AL-VA4(R)"        
#> [223] "AL-VL2p(R)"        "dACA(R)"           "a'L(L)"           
#> [226] "CX"                "AOTU(R)"           "ME(R)"            
#> [229] "AL-DA3(R)"         "AL-VA3(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method     path                                           
#> [1,] "GET"      "/api"                                         
#> [2,] "GET"      "/api/dbmeta/database"                         
#> [3,] "GET"      "/api/npexplorer/neuronmeta"                   
#> [4,] "POST"     "/api/v:ver/npexplorer/roiconnectivity"        
#> [5,] "GET"      "/api/v:ver/npexplorer/celltype/:dataset/:type"
#> [6,] "PROPFIND" "/api"                                         
#>      name                                                                                      
#> [1,] "github.com/labstack/echo.glob..func1"                                                    
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatabase-fm"          
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMeta-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                      
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                      
#> [6,] "github.com/labstack/echo.glob..func1"                                                    
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }