Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "CRE(R)"            "FBl7"              "IB"               
#>   [4] "PED(R)"            "WED(R)"            "PVLP(R)"          
#>   [7] "FBl6"              "PB(R5)"            "AB(L)"            
#>  [10] "PB(L2)"            "PB(R6)"            "LX(R)"            
#>  [13] "b'L(L)"            "FBl8"              "vACA(R)"          
#>  [16] "LAL(L)"            "FLA(R)"            "aL(R)"            
#>  [19] "AOTU(R)"           "SIP(L)"            "PB(R3)"           
#>  [22] "a1(R)"             "NO2(L)"            "MB(L)"            
#>  [25] "PB(R9)"            "bL(R)"             "PB(L6)"           
#>  [28] "g5(R)"             "g1(R)"             "PB(R1)"           
#>  [31] "CAN(R)"            "PB(L3)"            "LAL(-GA)(R)"      
#>  [34] "PB(L1)"            "ME(R)"             "PB(L9)"           
#>  [37] "EBr3pw"            "ICL(R)"            "NO1(L)"           
#>  [40] "g2(R)"             "a'3(R)"            "a'2(R)"           
#>  [43] "EBr3d"             "aL(L)"             "GC"               
#>  [46] "LH(R)"             "PB(R4)"            "SLP(R)"           
#>  [49] "g3(R)"             "g4(R)"             "PB"               
#>  [52] "CRE(L)"            "a'1(R)"            "SAD"              
#>  [55] "SCL(R)"            "FB"                "EBr5"             
#>  [58] "VMNP"              "EB"                "SAD(-AMMC)"       
#>  [61] "NO2(R)"            "AL(R)"             "lACA(R)"          
#>  [64] "EBr6"              "SPS(L)"            "MB(R)"            
#>  [67] "LOP(R)"            "SMP(L)"            "BU(L)"            
#>  [70] "ICL(L)"            "LO(R)"             "b'1(R)"           
#>  [73] "PB(R7)"            "PB(L8)"            "EBr2r4"           
#>  [76] "SNP(L)"            "POC"               "EBr3am"           
#>  [79] "GA(R)"             "PRW"               "AME(R)"           
#>  [82] "INP"               "NO(L)"             "a'L(L)"           
#>  [85] "AOT(R)"            "a2(R)"             "bL(L)"            
#>  [88] "NO"                "NO3(L)"            "VES(R)"           
#>  [91] "AVLP(R)"           "VLNP(R)"           "CRE(-ROB,-RUB)(R)"
#>  [94] "PB(R8)"            "AL(L)"             "CA(L)"            
#>  [97] "OL(R)"             "VES(L)"            "mALT(R)"          
#> [100] "FBl5"              "GNG"               "LAL(R)"           
#> [103] "ATL(L)"            "a'L(R)"            "FBl9"             
#> [106] "FBl1"              "PB(R2)"            "dACA(R)"          
#> [109] "EPA(R)"            "NO(R)"             "CA(R)"            
#> [112] "LX(L)"             "mALT(L)"           "NO1(R)"           
#> [115] "SIP(R)"            "SCL(L)"            "a3(R)"            
#> [118] "SMP(R)"            "ATL(R)"            "PB(L5)"           
#> [121] "AMMC"              "EBr1"              "gL(L)"            
#> [124] "BU(R)"             "SPS(R)"            "b'L(R)"           
#> [127] "ROB(R)"            "FBl2"              "AB(R)"            
#> [130] "EPA(L)"            "PENP"              "b'2(R)"           
#> [133] "CX"                "RUB(L)"            "GOR(R)"           
#> [136] "gL(R)"             "FBl3"              "b1(R)"            
#> [139] "RUB(R)"            "GOR(L)"            "SNP(R)"           
#> [142] "PLP(R)"            "FBl4"              "PB(L4)"           
#> [145] "hemibrain"         "NO3(R)"            "PB(L7)"           
#> [148] "MB(+ACA)(R)"       "IPS(R)"            "GF(R)"            
#> [151] "b2(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                              
#> [1,] "GET"  "/api/dbmeta/instances"           
#> [2,] "POST" "/api/v:ver/npexplorer/neuronmeta"
#> [3,] "GET"  "/api/available"                  
#> [4,] "GET"  "/login"                          
#> [5,] "POST" "/logout"                         
#> [6,] "GET"  "/api/help/*"                     
#>      name                                                                                 
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDataInstances-fm"
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                 
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.SetupRoutes.func1"                  
#> [4,] "github.com/janelia-flyem/echo-secure.loginHandler"                                  
#> [5,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"             
#> [6,] "github.com/labstack/echo.common.static.func1"                                       

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }