Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "PB(R1)"            "bL(R)"             "AL-DL5(R)"        
#>   [4] "mALT(R)"           "FBl5"              "AL-DM4(R)"        
#>   [7] "AL-VP1d(R)"        "AL-D(R)"           "AL-VA3(R)"        
#>  [10] "WED(R)"            "EB"                "AL-DA4m(L)"       
#>  [13] "PB(L6)"            "LX(L)"             "PB(R4)"           
#>  [16] "GF(R)"             "SMP(L)"            "lACA(R)"          
#>  [19] "FBl7"              "AL-DM3(L)"         "AL-DM4(L)"        
#>  [22] "AL-VA6(L)"         "NO"                "AL-DL4(L)"        
#>  [25] "PB(R2)"            "AL(R)"             "AL-VA2(R)"        
#>  [28] "AL-VM2(R)"         "IB"                "g1(R)"            
#>  [31] "PB(L7)"            "AL-VC3l(R)"        "VMNP"             
#>  [34] "AL-DM2(L)"         "VES(L)"            "SLP(R)"           
#>  [37] "AL-DM2(R)"         "AL-VL2p(R)"        "PB(L2)"           
#>  [40] "CRE(-ROB,-RUB)(R)" "b'2(R)"            "AOTU(R)"          
#>  [43] "AL-DC1(L)"         "bL(L)"             "NO(L)"            
#>  [46] "AL-VA5(R)"         "AL-DA4m(R)"        "EBr5"             
#>  [49] "FLA(R)"            "CRE(L)"            "PB(R7)"           
#>  [52] "ICL(L)"            "EPA(R)"            "AL-DP1m(L)"       
#>  [55] "IPS(R)"            "AL-DA1(R)"         "LAL(-GA)(R)"      
#>  [58] "PB(L9)"            "ATL(R)"            "AL-DL1(R)"        
#>  [61] "BU(R)"             "NO3(L)"            "SAD"              
#>  [64] "IB112"             "AL-V(R)"           "VLNP(R)"          
#>  [67] "dACA(R)"           "AL-VM5d(R)"        "PED(R)"           
#>  [70] "SCL(R)"            "a'1(R)"            "INP"              
#>  [73] "LAL(R)"            "AL-DA4l(R)"        "PENP"             
#>  [76] "EBr2r4"            "NO2(R)"            "AL-VM4(R)"        
#>  [79] "hemibrain"         "aL(R)"             "EBr1"             
#>  [82] "AL-DL5(L)"         "FBl3"              "GOR(R)"           
#>  [85] "AL-VA1v(R)"        "AL-VA6(R)"         "BU(L)"            
#>  [88] "ME(R)"             "ROB(R)"            "EBr3pw"           
#>  [91] "AL-VM7v(R)"        "RUB(R)"            "PB(L5)"           
#>  [94] "CRE(R)"            "SNP(L)"            "AL-DM6(R)"        
#>  [97] "AL-VM1(R)"         "AL-VM5v(R)"        "SCL(L)"           
#> [100] "g4(R)"             "CRE(-RUB)(L)"      "a3(R)"            
#> [103] "SIP(R)"            "vACA(R)"           "AL-DC4(L)"        
#> [106] "PB(L3)"            "PB(L4)"            "a1(R)"            
#> [109] "GC"                "gL(L)"             "NO(R)"            
#> [112] "VES(R)"            "AL-DA3(R)"         "CA(L)"            
#> [115] "FB"                "PLP(R)"            "a'2(R)"           
#> [118] "AL-VA1d(R)"        "AL-VP3(R)"         "NO3(R)"           
#> [121] "AL(L)"             "PB(R3)"            "PB(L8)"           
#> [124] "FBl2"              "CX"                "AL-D(L)"          
#> [127] "gL(R)"             "b1(R)"             "AL-DC2(R)"        
#> [130] "AL-DL4(R)"         "PB(R6)"            "SIP(L)"           
#> [133] "SPS(L)"            "AL-DM1(L)"         "AB(R)"            
#> [136] "AL-DP1m(R)"        "EBr6"              "FBl8"             
#> [139] "ICL(R)"            "AL-DM5(R)"         "AL-VM7d(R)"       
#> [142] "AME(R)"            "AL-DM1(R)"         "AOT(R)"           
#> [145] "b'L(L)"            "AL-VP5(R)"         "AL-VP1l(R)"       
#> [148] "b'1(R)"            "AMMC"              "FBl6"             
#> [151] "AL-VC3m(R)"        "AL-VL2a(R)"        "AL-VP4(R)"        
#> [154] "LOP(R)"            "SPS(R)"            "b2(R)"            
#> [157] "g2(R)"             "RUB(L)"            "AL-DA3(L)"        
#> [160] "GNG"               "a'L(L)"            "AL-DM3(R)"        
#> [163] "POC"               "EPA(L)"            "AL-VA4(R)"        
#> [166] "PB(R5)"            "AL-VP1m(R)"        "NO2(L)"           
#> [169] "g3(R)"             "AL-DP1l(R)"        "AL-VM3(R)"        
#> [172] "AL-VL1(R)"         "AB(L)"             "EBr3d"            
#> [175] "LO(R)"             "AVLP(R)"           "CAN(R)"           
#> [178] "AL-DA2(R)"         "AL-DC1(R)"         "SMP(R)"           
#> [181] "GOR(L)"            "AL-VA7m(R)"        "AL-VC2(R)"        
#> [184] "a'L(R)"            "AL-DC3(R)"         "a'3(R)"           
#> [187] "AL-VA7l(R)"        "PRW"               "b'L(R)"           
#> [190] "NO1(R)"            "AL-DM5(L)"         "MB(+ACA)(R)"      
#> [193] "FBl9"              "g5(R)"             "aL(L)"            
#> [196] "GA(R)"             "LX(R)"             "SNP(R)"           
#> [199] "SAD(-AMMC)"        "AL-DL2v(R)"        "ATL(L)"           
#> [202] "AL-DC4(R)"         "NO1(L)"            "FBl1"             
#> [205] "LH(R)"             "OL(R)"             "EBr3am"           
#> [208] "AL-DM6(L)"         "AL-DL3(R)"         "AL-VC1(R)"        
#> [211] "PVLP(R)"           "PB(R8)"            "AL-DA2(L)"        
#> [214] "AL-DC2(L)"         "AL-VM7d(L)"        "FBl4"             
#> [217] "MB(R)"             "AL-DL2d(R)"        "mALT(L)"          
#> [220] "PB(L1)"            "AL-VM7v(L)"        "AL-VC5(R)"        
#> [223] "LAL(L)"            "MB(L)"             "PB(R9)"           
#> [226] "FB-column3"        "CA(R)"             "AL-VC4(R)"        
#> [229] "a2(R)"             "PB"                "AL-VP2(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "EB"                "NO"                "LH(R)"            
#>   [4] "g3(R)"             "GA(R)"             "PRW"              
#>   [7] "b'1(R)"            "FBl5"              "AVLP(R)"          
#>  [10] "LOP(R)"            "hemibrain"         "SIP(L)"           
#>  [13] "g1(R)"             "MB(L)"             "g2(R)"            
#>  [16] "IB"                "PB(L3)"            "a'1(R)"           
#>  [19] "SNP(L)"            "a3(R)"             "OL(R)"            
#>  [22] "VLNP(R)"           "PB(R8)"            "NO(R)"            
#>  [25] "b1(R)"             "RUB(L)"            "VES(L)"           
#>  [28] "PB(L2)"            "AOT(R)"            "a2(R)"            
#>  [31] "SPS(R)"            "PED(R)"            "NO3(R)"           
#>  [34] "EBr5"              "FBl7"              "bL(R)"            
#>  [37] "vACA(R)"           "SNP(R)"            "PB(R9)"           
#>  [40] "gL(R)"             "CAN(R)"            "PB(R7)"           
#>  [43] "GOR(L)"            "PB(L8)"            "g5(R)"            
#>  [46] "b'2(R)"            "LAL(L)"            "CRE(R)"           
#>  [49] "NO1(R)"            "PB(L1)"            "SIP(R)"           
#>  [52] "CRE(-ROB,-RUB)(R)" "SAD"               "GC"               
#>  [55] "FLA(R)"            "LX(R)"             "gL(L)"            
#>  [58] "FBl4"              "b'L(L)"            "NO(L)"            
#>  [61] "SMP(R)"            "EBr3pw"            "ATL(L)"           
#>  [64] "a1(R)"             "GF(R)"             "CRE(L)"           
#>  [67] "FBl2"              "LAL(R)"            "SMP(L)"           
#>  [70] "a'L(R)"            "ICL(L)"            "FBl1"             
#>  [73] "FBl6"              "CA(R)"             "EBr3am"           
#>  [76] "POC"               "AOTU(R)"           "b'L(R)"           
#>  [79] "BU(L)"             "EBr2r4"            "FBl3"             
#>  [82] "LX(L)"             "INP"               "ME(R)"            
#>  [85] "g4(R)"             "NO2(L)"            "dACA(R)"          
#>  [88] "MB(+ACA)(R)"       "FBl9"              "SCL(L)"           
#>  [91] "CA(L)"             "PENP"              "PB(R6)"           
#>  [94] "NO1(L)"            "PB(L6)"            "ATL(R)"           
#>  [97] "GNG"               "PB(L5)"            "SPS(L)"           
#> [100] "PB(L9)"            "PVLP(R)"           "AL(R)"            
#> [103] "EBr1"              "IPS(R)"            "BU(R)"            
#> [106] "NO2(R)"            "a'L(L)"            "PB(L4)"           
#> [109] "FB"                "PLP(R)"            "SAD(-AMMC)"       
#> [112] "FBl8"              "ICL(R)"            "AME(R)"           
#> [115] "GOR(R)"            "RUB(R)"            "mALT(R)"          
#> [118] "WED(R)"            "PB(R3)"            "PB(R5)"           
#> [121] "CX"                "SCL(R)"            "a'2(R)"           
#> [124] "AL(L)"             "bL(L)"             "PB(R2)"           
#> [127] "aL(L)"             "AMMC"              "EBr3d"            
#> [130] "PB(L7)"            "LO(R)"             "VES(R)"           
#> [133] "lACA(R)"           "aL(R)"             "VMNP"             
#> [136] "AB(R)"             "PB(R4)"            "b2(R)"            
#> [139] "EBr6"              "PB"                "NO3(L)"           
#> [142] "MB(R)"             "AB(L)"             "ROB(R)"           
#> [145] "EPA(L)"            "SLP(R)"            "a'3(R)"           
#> [148] "mALT(L)"           "LAL(-GA)(R)"       "EPA(R)"           
#> [151] "PB(R1)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "PB(L3)"            "SPS(R)"            "VMNP"             
#>   [4] "SLP(R)"            "bL(L)"             "ICL(L)"           
#>   [7] "AL-VM5v(R)"        "NO(L)"             "a'1(R)"           
#>  [10] "OL(R)"             "AL-DA3(R)"         "AL-DM6(R)"        
#>  [13] "AL-VL2p(R)"        "AL-V(R)"           "a2(R)"            
#>  [16] "aL(R)"             "AL-DC4(R)"         "hemibrain"        
#>  [19] "NO3(R)"            "FBl2"              "LAL(R)"           
#>  [22] "AL-DL4(R)"         "CRE(-ROB,-RUB)(R)" "CRE(R)"           
#>  [25] "AL-DA2(L)"         "PB(R9)"            "b'L(L)"           
#>  [28] "AL(R)"             "GOR(L)"            "LAL(-GA)(R)"      
#>  [31] "AL-VM7v(L)"        "ROB(R)"            "AL-VL2a(R)"       
#>  [34] "AOT(R)"            "SCL(L)"            "ATL(L)"           
#>  [37] "CA(L)"             "PLP(R)"            "EB"               
#>  [40] "AVLP(R)"           "b1(R)"             "GOR(R)"           
#>  [43] "LAL(L)"            "GF(R)"             "AL-DL5(L)"        
#>  [46] "FBl6"              "aL(L)"             "AL-VA6(R)"        
#>  [49] "IB"                "PB(L2)"            "SCL(R)"           
#>  [52] "EPA(L)"            "AL-DL1(R)"         "AL-DL2v(R)"       
#>  [55] "AL-DP1l(R)"        "a'L(R)"            "MB(+ACA)(R)"      
#>  [58] "g1(R)"             "PB(L9)"            "AL-VC3m(R)"       
#>  [61] "AL-VP1l(R)"        "PB(L1)"            "NO3(L)"           
#>  [64] "AL-VM5d(R)"        "AMMC"              "ICL(R)"           
#>  [67] "a3(R)"             "AL-VA3(R)"         "mALT(R)"          
#>  [70] "AL-DC2(L)"         "FBl1"              "AL-DL5(R)"        
#>  [73] "AL-VP1d(R)"        "PB(R8)"            "GNG"              
#>  [76] "AL-VM7d(R)"        "PB(R4)"            "CX"               
#>  [79] "NO2(L)"            "AME(R)"            "AL-DM3(R)"        
#>  [82] "AL-VA1v(R)"        "PENP"              "b'2(R)"           
#>  [85] "AL-VM7v(R)"        "PB(R6)"            "FB"               
#>  [88] "AL-DM5(L)"         "AL-DA1(R)"         "AL-VL1(R)"        
#>  [91] "SMP(L)"            "dACA(R)"           "EBr3d"            
#>  [94] "AL-DA4m(R)"        "EBr6"              "EPA(R)"           
#>  [97] "LOP(R)"            "NO"                "FBl8"             
#> [100] "ATL(R)"            "b2(R)"             "gL(L)"            
#> [103] "PB(L5)"            "g4(R)"             "CRE(-RUB)(L)"     
#> [106] "LO(R)"             "PVLP(R)"           "WED(R)"           
#> [109] "BU(L)"             "BU(R)"             "AL-DM3(L)"        
#> [112] "gL(R)"             "FBl7"              "MB(R)"            
#> [115] "GA(R)"             "SIP(L)"            "NO(R)"            
#> [118] "vACA(R)"           "a1(R)"             "AL-DA2(R)"        
#> [121] "g5(R)"             "LH(R)"             "AL-VA1d(R)"       
#> [124] "SNP(L)"            "PB(R3)"            "LX(L)"            
#> [127] "VES(R)"            "AL(L)"             "FBl4"             
#> [130] "AL-DL3(R)"         "MB(L)"             "a'2(R)"           
#> [133] "FBl9"              "AL-VP2(R)"         "mALT(L)"          
#> [136] "GC"                "g2(R)"             "PB(L8)"           
#> [139] "EBr2r4"            "CRE(L)"            "a'L(L)"           
#> [142] "RUB(R)"            "AL-DL4(L)"         "VES(L)"           
#> [145] "AL-DM2(R)"         "AL-DM4(R)"         "AL-VA5(R)"        
#> [148] "g3(R)"             "VLNP(R)"           "NO1(L)"           
#> [151] "PB(L6)"            "AL-DA4l(R)"        "AL-D(R)"          
#> [154] "AL-VA4(R)"         "AL-VC3l(R)"        "PB(R7)"           
#> [157] "AL-VC2(R)"         "AOTU(R)"           "AL-DP1m(L)"       
#> [160] "PB(R5)"            "AL-VP4(R)"         "AL-VM4(R)"        
#> [163] "b'L(R)"            "EBr5"              "AL-DM4(L)"        
#> [166] "FLA(R)"            "EBr3pw"            "AL-VC4(R)"        
#> [169] "AL-DC3(R)"         "EBr1"              "SPS(L)"           
#> [172] "PB"                "SIP(R)"            "PB(R1)"           
#> [175] "PB(L4)"            "AL-DM1(R)"         "AL-DM5(R)"        
#> [178] "AL-VC1(R)"         "PB(R2)"            "CAN(R)"           
#> [181] "POC"               "AL-VA7l(R)"        "AL-VM3(R)"        
#> [184] "AL-VP5(R)"         "AL-VP3(R)"         "lACA(R)"          
#> [187] "SNP(R)"            "AL-DM1(L)"         "AL-DM2(L)"        
#> [190] "IPS(R)"            "NO2(R)"            "AL-DC2(R)"        
#> [193] "PB(L7)"            "FBl3"              "AL-DC1(L)"        
#> [196] "FBl5"              "bL(R)"             "a'3(R)"           
#> [199] "NO1(R)"            "ME(R)"             "AL-DC4(L)"        
#> [202] "AL-VA6(L)"         "FB-column3"        "INP"              
#> [205] "AL-VC5(R)"         "AL-VM1(R)"         "EBr3am"           
#> [208] "SAD"               "AL-VM7d(L)"        "AL-VA2(R)"        
#> [211] "AL-DA4m(L)"        "AB(L)"             "PED(R)"           
#> [214] "AB(R)"             "RUB(L)"            "AL-DP1m(R)"       
#> [217] "AL-VM2(R)"         "AL-VA7m(R)"        "AL-D(L)"          
#> [220] "AL-DL2d(R)"        "AL-VP1m(R)"        "SMP(R)"           
#> [223] "LX(R)"             "AL-DM6(L)"         "AL-DC1(R)"        
#> [226] "PRW"               "AL-DA3(L)"         "SAD(-AMMC)"       
#> [229] "CA(R)"             "b'1(R)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "GC"                "AL-DL5(L)"         "a'1(R)"           
#>   [4] "RUB(R)"            "CA(L)"             "AME(R)"           
#>   [7] "AL-DM5(R)"         "PB(R8)"            "a'L(L)"           
#>  [10] "NO3(R)"            "NO(R)"             "gL(R)"            
#>  [13] "ICL(R)"            "AL-DA3(R)"         "AL-VA5(R)"        
#>  [16] "NO1(L)"            "b2(R)"             "AL-DL4(L)"        
#>  [19] "FBl5"              "CA(R)"             "GNG"              
#>  [22] "AL-DL4(R)"         "a1(R)"             "RUB(L)"           
#>  [25] "LAL(L)"            "PB(L1)"            "SPS(L)"           
#>  [28] "VMNP"              "AL-DM2(L)"         "IPS(R)"           
#>  [31] "dACA(R)"           "AL-DA4m(L)"        "MB(R)"            
#>  [34] "IB"                "aL(R)"             "SNP(R)"           
#>  [37] "AL-VM7v(L)"        "AL-DL2d(R)"        "AL-VA2(R)"        
#>  [40] "LOP(R)"            "SAD(-AMMC)"        "LO(R)"            
#>  [43] "MB(+ACA)(R)"       "bL(R)"             "AL-DC3(R)"        
#>  [46] "AOT(R)"            "CX"                "PENP"             
#>  [49] "hemibrain"         "PB(R9)"            "PB(L6)"           
#>  [52] "NO2(R)"            "b'2(R)"            "PRW"              
#>  [55] "EPA(R)"            "aL(L)"             "AL-DC2(R)"        
#>  [58] "VLNP(R)"           "AL-DC4(L)"         "CRE(L)"           
#>  [61] "AL-VM5d(R)"        "EBr3am"            "AL-DA2(L)"        
#>  [64] "b'L(L)"            "AL(R)"             "AL-VM2(R)"        
#>  [67] "a2(R)"             "GOR(L)"            "FLA(R)"           
#>  [70] "BU(R)"             "NO"                "a'L(R)"           
#>  [73] "AL-DC1(L)"         "AL-DC2(L)"         "ATL(R)"           
#>  [76] "AVLP(R)"           "SLP(R)"            "LAL(-GA)(R)"      
#>  [79] "AB(R)"             "AL-DM3(R)"         "AL-VP5(R)"        
#>  [82] "GF(R)"             "AL-DM3(L)"         "PVLP(R)"          
#>  [85] "FBl3"              "LAL(R)"            "MB(L)"            
#>  [88] "a'2(R)"            "AL-VA1v(R)"        "AL-VA7l(R)"       
#>  [91] "gL(L)"             "g5(R)"             "AL-D(R)"          
#>  [94] "AL-VL1(R)"         "PLP(R)"            "ME(R)"            
#>  [97] "AL-D(L)"           "SIP(R)"            "NO3(L)"           
#> [100] "EBr3pw"            "AL(L)"             "AL-DM4(L)"        
#> [103] "AL-DM6(L)"         "FBl8"              "FBl9"             
#> [106] "AL-DL1(R)"         "AL-VM3(R)"         "g1(R)"            
#> [109] "EBr6"              "NO1(R)"            "AL-DC1(R)"        
#> [112] "mALT(L)"           "ROB(R)"            "AL-DM4(R)"        
#> [115] "AL-VP1l(R)"        "AL-VA7m(R)"        "AL-VC1(R)"        
#> [118] "AL-DP1m(L)"        "AL-VA1d(R)"        "SNP(L)"           
#> [121] "FB-column3"        "AL-DA4m(R)"        "AL-VA4(R)"        
#> [124] "GA(R)"             "AL-DM6(R)"         "AL-VL2a(R)"       
#> [127] "ATL(L)"            "BU(L)"             "AL-DM1(L)"        
#> [130] "AL-VC4(R)"         "PB(R6)"            "g3(R)"            
#> [133] "PB(L3)"            "SCL(R)"            "FB"               
#> [136] "LX(R)"             "PB(L7)"            "FBl6"             
#> [139] "AL-VC3m(R)"        "bL(L)"             "PB(R3)"           
#> [142] "INP"               "AL-VM4(R)"         "b'L(R)"           
#> [145] "VES(L)"            "AL-DA4l(R)"        "AL-VC2(R)"        
#> [148] "a'3(R)"            "EBr5"              "AL-DA3(L)"        
#> [151] "AL-VP3(R)"         "vACA(R)"           "PB"               
#> [154] "AL-VM7d(L)"        "AL-VL2p(R)"        "AL-VM7v(R)"       
#> [157] "AL-VP4(R)"         "AL-DM5(L)"         "NO(L)"            
#> [160] "AL-VA3(R)"         "PB(R2)"            "AL-DA2(R)"        
#> [163] "AL-DP1l(R)"        "AL-VC5(R)"         "mALT(R)"          
#> [166] "CRE(R)"            "a3(R)"             "AL-DA1(R)"        
#> [169] "AL-VP2(R)"         "SMP(L)"            "EPA(L)"           
#> [172] "PB(L2)"            "LH(R)"             "SCL(L)"           
#> [175] "PB(R7)"            "PED(R)"            "WED(R)"           
#> [178] "PB(L9)"            "CRE(-RUB)(L)"      "CAN(R)"           
#> [181] "PB(R4)"            "PB(L4)"            "AL-V(R)"          
#> [184] "lACA(R)"           "EBr2r4"            "FBl2"             
#> [187] "LX(L)"             "GOR(R)"            "AL-VM7d(R)"       
#> [190] "POC"               "PB(L5)"            "g4(R)"            
#> [193] "AL-VP1m(R)"        "AL-VA6(L)"         "AL-DM1(R)"        
#> [196] "SAD"               "VES(R)"            "g2(R)"            
#> [199] "AL-VC3l(R)"        "AOTU(R)"           "SPS(R)"           
#> [202] "EBr1"              "AL-DL5(R)"         "AL-VM1(R)"        
#> [205] "AL-VP1d(R)"        "AB(L)"             "NO2(L)"           
#> [208] "CRE(-ROB,-RUB)(R)" "SIP(L)"            "PB(L8)"           
#> [211] "FBl4"              "b1(R)"             "AL-DL3(R)"        
#> [214] "b'1(R)"            "AMMC"              "FBl1"             
#> [217] "AL-DM2(R)"         "AL-VA6(R)"         "AL-DC4(R)"        
#> [220] "AL-DL2v(R)"        "PB(R5)"            "EBr3d"            
#> [223] "ICL(L)"            "AL-VM5v(R)"        "SMP(R)"           
#> [226] "EB"                "PB(R1)"            "FBl7"             
#> [229] "OL(R)"             "AL-DP1m(R)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                          
#> [1,] "POST" "/api"                                        
#> [2,] "POST" "/api/v:ver/raw/cypher/transaction/:id/commit"
#> [3,] "PUT"  "/api"                                        
#> [4,] "POST" "/api/v:ver/npexplorer/rankedtable"           
#> [5,] "POST" "/api/v:ver/npexplorer/roisinneuron"          
#> [6,] "GET"  "/api/npexplorer/completeness"                
#>      name                                                                                        
#> [1,] "github.com/labstack/echo.glob..func1"                                                      
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [3,] "github.com/labstack/echo.glob..func1"                                                      
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCompleteness-fm"
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }