Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "VES(R)"            "FBl7"              "AL-DA3(R)"        
#>   [4] "AL-DC1(R)"         "AOT(R)"            "PB(L5)"           
#>   [7] "AL-VA1d(R)"        "BU(L)"             "EB"               
#>  [10] "AL-D(L)"           "AL-VC2(R)"         "AL-VC4(R)"        
#>  [13] "EBr3am"            "EPA(R)"            "NO2(L)"           
#>  [16] "PLP(R)"            "a'2(R)"            "AL-DL4(L)"        
#>  [19] "SLP(R)"            "AL-VM3(R)"         "AL-V(R)"          
#>  [22] "CX"                "PB(R2)"            "AL-VA1v(R)"       
#>  [25] "IB"                "PB(R6)"            "CA(R)"            
#>  [28] "AL-DA4l(R)"        "AL-VP1d(R)"        "b'1(R)"           
#>  [31] "AL-VA5(R)"         "AL-DA4m(L)"        "AL-DM4(L)"        
#>  [34] "NO(R)"             "FBl2"              "AL-VC3m(R)"       
#>  [37] "CA(L)"             "PB(R7)"            "AL-DC1(L)"        
#>  [40] "PB(L9)"            "EBr1"              "AL-VP5(R)"        
#>  [43] "AL-VP3(R)"         "GOR(L)"            "b2(R)"            
#>  [46] "FBl5"              "PB(R4)"            "a'L(L)"           
#>  [49] "mALT(L)"           "PB(L8)"            "AL(L)"            
#>  [52] "AL-DA3(L)"         "AL-DM3(L)"         "AL(R)"            
#>  [55] "AL-DL4(R)"         "AL-DM1(R)"         "AL-DM3(R)"        
#>  [58] "AL-DP1m(R)"        "PB(R8)"            "a'L(R)"           
#>  [61] "AL-DC4(L)"         "dACA(R)"           "PB(L7)"           
#>  [64] "OL(R)"             "AL-DP1l(R)"        "vACA(R)"          
#>  [67] "LX(R)"             "SNP(R)"            "FB-column3"       
#>  [70] "FBl3"              "PB(R5)"            "SIP(L)"           
#>  [73] "b'L(R)"            "g4(R)"             "SPS(R)"           
#>  [76] "NO1(L)"            "PB"                "IB112"            
#>  [79] "LH(R)"             "SNP(L)"            "IPS(R)"           
#>  [82] "EPA(L)"            "AL-DC3(R)"         "AL-VM7d(R)"       
#>  [85] "a1(R)"             "PENP"              "a2(R)"            
#>  [88] "AME(R)"            "NO2(R)"            "AL-DA1(R)"        
#>  [91] "AL-VA4(R)"         "AL-VL1(R)"         "GA(R)"            
#>  [94] "NO3(R)"            "RUB(L)"            "AL-VA7m(R)"       
#>  [97] "AL-VP1m(R)"        "PRW"               "SIP(R)"           
#> [100] "SCL(L)"            "LAL(-GA)(R)"       "aL(R)"            
#> [103] "AL-VA6(R)"         "SCL(R)"            "AL-DL5(L)"        
#> [106] "NO(L)"             "MB(R)"             "AL-DM2(R)"        
#> [109] "AL-DM4(R)"         "AL-DM5(R)"         "NO3(L)"           
#> [112] "AL-DL5(R)"         "AL-VA2(R)"         "AL-VC1(R)"        
#> [115] "ICL(R)"            "VMNP"              "PB(L6)"           
#> [118] "AL-DL1(R)"         "AL-VP4(R)"         "g2(R)"            
#> [121] "BU(R)"             "bL(L)"             "SAD(-AMMC)"       
#> [124] "AB(L)"             "a'3(R)"            "AL-DL3(R)"        
#> [127] "AOTU(R)"           "EBr6"              "FBl4"             
#> [130] "LX(L)"             "AL-VA3(R)"         "PED(R)"           
#> [133] "FBl9"              "CAN(R)"            "RUB(R)"           
#> [136] "AL-DC2(L)"         "AL-VM7v(L)"        "SPS(L)"           
#> [139] "EBr3d"             "ATL(R)"            "AL-VM1(R)"        
#> [142] "AL-VP1l(R)"        "gL(L)"             "ICL(L)"           
#> [145] "AL-DC4(R)"         "SAD"               "g5(R)"            
#> [148] "AL-DC2(R)"         "LOP(R)"            "hemibrain"        
#> [151] "AL-DA2(L)"         "gL(R)"             "WED(R)"           
#> [154] "CRE(R)"            "AL-VM4(R)"         "AL-VP2(R)"        
#> [157] "SMP(R)"            "AL-VM7d(L)"        "b1(R)"            
#> [160] "AL-VM7v(R)"        "ATL(L)"            "GC"               
#> [163] "AL-VC5(R)"         "mALT(R)"           "a'1(R)"           
#> [166] "PVLP(R)"           "LAL(R)"            "AL-VM5v(R)"       
#> [169] "GF(R)"             "GOR(R)"            "AL-DM2(L)"        
#> [172] "FBl1"              "AL-VA6(L)"         "PB(R9)"           
#> [175] "FBl6"              "AL-DA4m(R)"        "AL-D(R)"          
#> [178] "AL-VA7l(R)"        "b'2(R)"            "g1(R)"            
#> [181] "AB(R)"             "b'L(L)"            "AL-VM5d(R)"       
#> [184] "VLNP(R)"           "LAL(L)"            "PB(L3)"           
#> [187] "lACA(R)"           "AL-DM1(L)"         "POC"              
#> [190] "SMP(L)"            "FB"                "EBr3pw"           
#> [193] "aL(L)"             "AL-VC3l(R)"        "AL-VM2(R)"        
#> [196] "EBr2r4"            "GNG"               "AL-VL2p(R)"       
#> [199] "FLA(R)"            "AL-DP1m(L)"        "NO"               
#> [202] "FBl8"              "INP"               "AL-VL2a(R)"       
#> [205] "ME(R)"             "MB(+ACA)(R)"       "AMMC"             
#> [208] "ROB(R)"            "a3(R)"             "PB(L4)"           
#> [211] "AL-DL2d(R)"        "MB(L)"             "EBr5"             
#> [214] "bL(R)"             "CRE(-ROB,-RUB)(R)" "g3(R)"            
#> [217] "NO1(R)"            "CRE(L)"            "PB(R1)"           
#> [220] "AL-DL2v(R)"        "PB(L1)"            "PB(R3)"           
#> [223] "AL-DA2(R)"         "AL-DM6(L)"         "VES(L)"           
#> [226] "AL-DM6(R)"         "PB(L2)"            "AL-DM5(L)"        
#> [229] "CRE(-RUB)(L)"      "LO(R)"             "AVLP(R)"          
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "gL(L)"             "LO(R)"             "PVLP(R)"          
#>   [4] "SCL(L)"            "NO3(R)"            "a'L(R)"           
#>   [7] "MB(+ACA)(R)"       "ICL(R)"            "a'L(L)"           
#>  [10] "VES(L)"            "PRW"               "lACA(R)"          
#>  [13] "PB(R9)"            "POC"               "SPS(L)"           
#>  [16] "PLP(R)"            "CAN(R)"            "PB(R5)"           
#>  [19] "FBl1"              "LH(R)"             "SCL(R)"           
#>  [22] "ATL(R)"            "AME(R)"            "AOT(R)"           
#>  [25] "LOP(R)"            "PED(R)"            "GC"               
#>  [28] "EB"                "VES(R)"            "FBl9"             
#>  [31] "b'L(L)"            "PB(L3)"            "GA(R)"            
#>  [34] "AB(R)"             "mALT(L)"           "a1(R)"            
#>  [37] "AOTU(R)"           "SIP(R)"            "GNG"              
#>  [40] "FBl4"              "AMMC"              "BU(L)"            
#>  [43] "PB(L8)"            "SNP(R)"            "PB(L9)"           
#>  [46] "SMP(L)"            "hemibrain"         "FBl8"             
#>  [49] "b'1(R)"            "ATL(L)"            "SPS(R)"           
#>  [52] "PB(R2)"            "aL(L)"             "INP"              
#>  [55] "PB(L4)"            "VLNP(R)"           "SIP(L)"           
#>  [58] "mALT(R)"           "WED(R)"            "g1(R)"            
#>  [61] "PB(L1)"            "g5(R)"             "VMNP"             
#>  [64] "NO(R)"             "bL(L)"             "AVLP(R)"          
#>  [67] "EPA(L)"            "NO(L)"             "LX(R)"            
#>  [70] "CRE(L)"            "EBr3am"            "NO1(R)"           
#>  [73] "EBr3d"             "PB(R4)"            "OL(R)"            
#>  [76] "AB(L)"             "LX(L)"             "RUB(L)"           
#>  [79] "LAL(-GA)(R)"       "FB"                "EBr1"             
#>  [82] "EPA(R)"            "PB(R1)"            "FBl5"             
#>  [85] "a3(R)"             "EBr6"              "GOR(R)"           
#>  [88] "CA(L)"             "IPS(R)"            "NO2(R)"           
#>  [91] "PENP"              "b'L(R)"            "EBr5"             
#>  [94] "PB(R8)"            "FBl2"              "PB(R3)"           
#>  [97] "FBl3"              "SLP(R)"            "PB(L5)"           
#> [100] "dACA(R)"           "AL(L)"             "EBr3pw"           
#> [103] "b1(R)"             "CX"                "NO2(L)"           
#> [106] "SMP(R)"            "IB"                "aL(R)"            
#> [109] "CRE(R)"            "b2(R)"             "a'2(R)"           
#> [112] "NO3(L)"            "FBl6"              "PB(R6)"           
#> [115] "NO1(L)"            "LAL(L)"            "g2(R)"            
#> [118] "FLA(R)"            "PB(L6)"            "RUB(R)"           
#> [121] "a'3(R)"            "PB(L2)"            "vACA(R)"          
#> [124] "PB(R7)"            "EBr2r4"            "g4(R)"            
#> [127] "SAD(-AMMC)"        "MB(R)"             "AL(R)"            
#> [130] "b'2(R)"            "a2(R)"             "ME(R)"            
#> [133] "ROB(R)"            "g3(R)"             "MB(L)"            
#> [136] "BU(R)"             "a'1(R)"            "SNP(L)"           
#> [139] "bL(R)"             "PB"                "NO"               
#> [142] "FBl7"              "LAL(R)"            "GF(R)"            
#> [145] "GOR(L)"            "ICL(L)"            "PB(L7)"           
#> [148] "SAD"               "gL(R)"             "CA(R)"            
#> [151] "CRE(-ROB,-RUB)(R)"
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "b'2(R)"            "b'L(R)"            "AL-DA4m(R)"       
#>   [4] "VLNP(R)"           "PB(R5)"            "AOTU(R)"          
#>   [7] "hemibrain"         "AL-VA6(L)"         "PB(R2)"           
#>  [10] "b1(R)"             "LX(L)"             "AL-VM7d(R)"       
#>  [13] "RUB(R)"            "AL-DA3(R)"         "SCL(L)"           
#>  [16] "GC"                "EB"                "AL-VL2p(R)"       
#>  [19] "AL-VP1m(R)"        "PB(R6)"            "AL-VA7m(R)"       
#>  [22] "PB(L2)"            "PVLP(R)"           "FBl9"             
#>  [25] "AL-DA2(R)"         "AL-DM1(R)"         "PB"               
#>  [28] "PB(R4)"            "PB(L3)"            "LAL(-GA)(R)"      
#>  [31] "AL-DM4(L)"         "PB(R1)"            "EBr3pw"           
#>  [34] "EPA(R)"            "LO(R)"             "FBl4"             
#>  [37] "CA(R)"             "AL-DM2(R)"         "a2(R)"            
#>  [40] "FLA(R)"            "NO1(R)"            "LH(R)"            
#>  [43] "mALT(R)"           "PLP(R)"            "AL-DA2(L)"        
#>  [46] "AL-DM2(L)"         "b'1(R)"            "aL(R)"            
#>  [49] "AL-DL1(R)"         "SPS(R)"            "FBl6"             
#>  [52] "AL-VA5(R)"         "FBl7"              "ICL(L)"           
#>  [55] "AL-DA3(L)"         "ICL(R)"            "bL(R)"            
#>  [58] "AL-VP5(R)"         "EBr2r4"            "aL(L)"            
#>  [61] "INP"               "AL-VA4(R)"         "AL-VC1(R)"        
#>  [64] "PB(R7)"            "SNP(R)"            "PB(L7)"           
#>  [67] "AL-VA1d(R)"        "g4(R)"             "AL-VC3m(R)"       
#>  [70] "CA(L)"             "PRW"               "BU(L)"            
#>  [73] "SPS(L)"            "AL-DM5(L)"         "MB(+ACA)(R)"      
#>  [76] "PB(L8)"            "AL-DP1m(L)"        "AL-DC4(R)"        
#>  [79] "a'3(R)"            "AL-DM3(L)"         "SMP(R)"           
#>  [82] "gL(L)"             "AL-VA2(R)"         "AL-VC5(R)"        
#>  [85] "mALT(L)"           "EBr3d"             "AL-DC1(L)"        
#>  [88] "AL-DL4(L)"         "gL(R)"             "AL-V(R)"          
#>  [91] "AL-DC2(L)"         "AL-D(L)"           "EBr3am"           
#>  [94] "GOR(L)"            "OL(R)"             "AL-DA1(R)"        
#>  [97] "AL-DM4(R)"         "GA(R)"             "AL-VM7v(L)"       
#> [100] "bL(L)"             "FBl5"              "WED(R)"           
#> [103] "g1(R)"             "b2(R)"             "PB(L6)"           
#> [106] "FB-column3"        "IPS(R)"            "AL-DL2d(R)"       
#> [109] "AL-DL3(R)"         "AL-DP1l(R)"        "IB"               
#> [112] "SAD"               "CRE(R)"            "ROB(R)"           
#> [115] "NO2(R)"            "RUB(L)"            "AL-DC3(R)"        
#> [118] "AL-VM1(R)"         "ATL(L)"            "a'1(R)"           
#> [121] "AME(R)"            "EPA(L)"            "AL-VM5d(R)"       
#> [124] "AL-VP3(R)"         "SMP(L)"            "FBl8"             
#> [127] "AL-VA6(R)"         "AL-VC3l(R)"        "NO(R)"            
#> [130] "g5(R)"             "SAD(-AMMC)"        "GOR(R)"           
#> [133] "FBl2"              "AL-VM4(R)"         "AL-VM7d(L)"       
#> [136] "AL-VA3(R)"         "PB(L5)"            "MB(L)"            
#> [139] "a'2(R)"            "CRE(-RUB)(L)"      "ATL(R)"           
#> [142] "AL-DC2(R)"         "AL-VP2(R)"         "AL-VP1l(R)"       
#> [145] "ME(R)"             "GNG"               "PB(L4)"           
#> [148] "POC"               "PB(L1)"            "LX(R)"            
#> [151] "AL-VM3(R)"         "AL-VA1v(R)"        "NO(L)"            
#> [154] "AL(R)"             "NO1(L)"            "AL(L)"            
#> [157] "NO"                "a'L(L)"            "AL-DC1(R)"        
#> [160] "AL-DP1m(R)"        "AL-VC4(R)"         "BU(R)"            
#> [163] "NO2(L)"            "EBr6"              "SNP(L)"           
#> [166] "AL-DM3(R)"         "a1(R)"             "AL-DC4(L)"        
#> [169] "a3(R)"             "SLP(R)"            "VMNP"             
#> [172] "FBl3"              "MB(R)"             "AL-VA7l(R)"       
#> [175] "AB(L)"             "LAL(L)"            "VES(L)"           
#> [178] "AL-VP1d(R)"        "AB(R)"             "AL-D(R)"          
#> [181] "AL-VL1(R)"         "AL-VL2a(R)"        "SIP(L)"           
#> [184] "LOP(R)"            "CX"                "PENP"             
#> [187] "g3(R)"             "g2(R)"             "EBr5"             
#> [190] "NO3(L)"            "FBl1"              "AL-VM2(R)"        
#> [193] "lACA(R)"           "dACA(R)"           "AL-DM1(L)"        
#> [196] "AVLP(R)"           "b'L(L)"            "AL-DM5(R)"        
#> [199] "GF(R)"             "AL-DA4m(L)"        "AL-DM6(L)"        
#> [202] "VES(R)"            "AL-DA4l(R)"        "AL-VC2(R)"        
#> [205] "AL-VP4(R)"         "AOT(R)"            "NO3(R)"           
#> [208] "SCL(R)"            "EBr1"              "a'L(R)"           
#> [211] "AL-DL2v(R)"        "AL-VM5v(R)"        "PED(R)"           
#> [214] "FB"                "AL-DL5(L)"         "CAN(R)"           
#> [217] "LAL(R)"            "AL-DL4(R)"         "vACA(R)"          
#> [220] "SIP(R)"            "PB(R9)"            "AL-VM7v(R)"       
#> [223] "AMMC"              "AL-DL5(R)"         "AL-DM6(R)"        
#> [226] "CRE(-ROB,-RUB)(R)" "PB(R8)"            "PB(L9)"           
#> [229] "CRE(L)"            "PB(R3)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.2`
#> $`hemibrain:v1.2`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2`$ROIs
#>   [1] "PB(L1)"            "FBl8"              "AL-VA5(R)"        
#>   [4] "VLNP(R)"           "LOP(R)"            "PB(R9)"           
#>   [7] "FBl1"              "CAN(R)"            "MB(R)"            
#>  [10] "AL-VM5v(R)"        "AL-VM7d(R)"        "BU(R)"            
#>  [13] "EPA(R)"            "AL-DM5(L)"         "AL-VA6(L)"        
#>  [16] "AL(R)"             "AL-VA2(R)"         "a'3(R)"           
#>  [19] "hemibrain"         "PB(R2)"            "NO3(R)"           
#>  [22] "AL-DC2(L)"         "PB(L6)"            "AB(R)"            
#>  [25] "EPA(L)"            "AL-VP2(R)"         "AOT(R)"           
#>  [28] "AL-VL1(R)"         "AL-VM7v(R)"        "SAD"              
#>  [31] "PB(L7)"            "FBl2"              "FBl3"             
#>  [34] "AL-DL3(R)"         "AL-DC1(L)"         "AL-DA2(L)"        
#>  [37] "AL-DM6(L)"         "CRE(-RUB)(L)"      "CA(R)"            
#>  [40] "PB(R3)"            "AL-DA2(R)"         "AL-DC3(R)"        
#>  [43] "AL-DL2v(R)"        "IB"                "PB(R4)"           
#>  [46] "AL-VP1d(R)"        "AMMC"              "SPS(R)"           
#>  [49] "CRE(R)"            "VES(L)"            "AL-VM2(R)"        
#>  [52] "AL-VC4(R)"         "GOR(L)"            "SAD(-AMMC)"       
#>  [55] "a3(R)"             "b'2(R)"            "AL-VM7d(L)"       
#>  [58] "ICL(R)"            "AME(R)"            "aL(L)"            
#>  [61] "g3(R)"             "AL-DA4m(L)"        "PB(L9)"           
#>  [64] "AL-DM4(R)"         "b'1(R)"            "AL-DM1(L)"        
#>  [67] "AL-DM4(L)"         "gL(R)"             "SIP(R)"           
#>  [70] "SLP(R)"            "AL-VP4(R)"         "NO1(R)"           
#>  [73] "a'1(R)"            "LO(R)"             "ATL(L)"           
#>  [76] "VMNP"              "FBl5"              "vACA(R)"          
#>  [79] "SMP(L)"            "EBr3am"            "SIP(L)"           
#>  [82] "POC"               "AB(L)"             "CRE(-ROB,-RUB)(R)"
#>  [85] "BU(L)"             "PB(L5)"            "AL-VM7v(L)"       
#>  [88] "AL-VA3(R)"         "AL(L)"             "AL-VM5d(R)"       
#>  [91] "NO2(L)"            "PRW"               "ROB(R)"           
#>  [94] "NO(L)"             "aL(R)"             "OL(R)"            
#>  [97] "AL-DL5(L)"         "NO3(L)"            "AL-DA1(R)"        
#> [100] "b'L(R)"            "AL-VA7m(R)"        "GC"               
#> [103] "EBr5"              "AL-VC2(R)"         "lACA(R)"          
#> [106] "FLA(R)"            "FBl6"              "GNG"              
#> [109] "PB(L4)"            "AL-VA4(R)"         "AL-VL2p(R)"       
#> [112] "AL-VP5(R)"         "PENP"              "LAL(-GA)(R)"      
#> [115] "SNP(R)"            "AL-DP1m(L)"        "AL-DA4m(R)"       
#> [118] "AL-VM3(R)"         "INP"               "b'L(L)"           
#> [121] "EBr6"              "PB(L8)"            "g4(R)"            
#> [124] "PB(L3)"            "ATL(R)"            "CX"               
#> [127] "GA(R)"             "a'L(R)"            "AL-DA3(L)"        
#> [130] "ICL(L)"            "AL-DL5(R)"         "AL-VM1(R)"        
#> [133] "VES(R)"            "FBl7"              "AL-DA3(R)"        
#> [136] "AL-DA4l(R)"        "SCL(L)"            "PB(L2)"           
#> [139] "FB"                "b1(R)"             "AL-DP1l(R)"       
#> [142] "EBr1"              "ME(R)"             "PB(R1)"           
#> [145] "LH(R)"             "AL-DL1(R)"         "AL-D(R)"          
#> [148] "AL-VA1v(R)"        "SMP(R)"            "a1(R)"            
#> [151] "gL(L)"             "AL-DM3(L)"         "AL-DM2(R)"        
#> [154] "AL-VM4(R)"         "AL-VP1m(R)"        "AL-D(L)"          
#> [157] "g5(R)"             "MB(+ACA)(R)"       "AL-VA7l(R)"       
#> [160] "AL-VC5(R)"         "PB(R5)"            "RUB(R)"           
#> [163] "CA(L)"             "LAL(L)"            "a2(R)"            
#> [166] "PED(R)"            "WED(R)"            "SNP(L)"           
#> [169] "FB-column3"        "NO2(R)"            "AL-DM5(R)"        
#> [172] "AOTU(R)"           "EBr3d"             "EBr2r4"           
#> [175] "FBl4"              "LAL(R)"            "AL-DM6(R)"        
#> [178] "AL-DC4(L)"         "AVLP(R)"           "AL-DC2(R)"        
#> [181] "AL-DC4(R)"         "AL-VC3l(R)"        "MB(L)"            
#> [184] "AL-DM2(L)"         "CRE(L)"            "bL(L)"            
#> [187] "bL(R)"             "dACA(R)"           "EB"               
#> [190] "PVLP(R)"           "LX(L)"             "AL-DM1(R)"        
#> [193] "mALT(R)"           "GOR(R)"            "AL-VP1l(R)"       
#> [196] "LX(R)"             "RUB(L)"            "AL-VL2a(R)"       
#> [199] "PB(R7)"            "AL-VC1(R)"         "NO(R)"            
#> [202] "AL-DL4(R)"         "NO1(L)"            "SCL(R)"           
#> [205] "PLP(R)"            "a'2(R)"            "AL-DL4(L)"        
#> [208] "NO"                "AL-V(R)"           "AL-VP3(R)"        
#> [211] "PB"                "FBl9"              "AL-DM3(R)"        
#> [214] "AL-DP1m(R)"        "AL-VC3m(R)"        "AL-VA6(R)"        
#> [217] "IPS(R)"            "a'L(L)"            "SPS(L)"           
#> [220] "EBr3pw"            "AL-DC1(R)"         "mALT(L)"          
#> [223] "PB(R6)"            "g2(R)"             "AL-DL2d(R)"       
#> [226] "GF(R)"             "g1(R)"             "b2(R)"            
#> [229] "PB(R8)"            "AL-VA1d(R)"       
#> 
#> $`hemibrain:v1.2`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method  path                                        
#> [1,] "PATCH" "/api"                                      
#> [2,] "GET"   "/api/custom/custom"                        
#> [3,] "POST"  "/api/npexplorer/autapses"                  
#> [4,] "POST"  "/api/v:ver/raw/cypher/transaction/:id/kill"
#> [5,] "GET"   "/api/dbmeta/version"                       
#> [6,] "GET"   "/api/v:ver/dbmeta/version"                 
#>      name                                                                                    
#> [1,] "github.com/labstack/echo.glob..func1"                                                  
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"         
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getVersion-fm"         
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }