Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-VA7l(R)"        "AL-VP5(R)"         "AL-VL1(R)"        
#>   [4] "GOR(L)"            "AL-DA2(L)"         "PB(L4)"           
#>   [7] "AL-DA4l(R)"        "POC"               "PB(L5)"           
#>  [10] "AMMC"              "EPA(R)"            "AL-DL4(L)"        
#>  [13] "AL-DL5(R)"         "AL-VP2(R)"         "PLP(R)"           
#>  [16] "GNG"               "PB(L7)"            "FBl9"             
#>  [19] "SCL(L)"            "PB(L6)"            "PB(R4)"           
#>  [22] "OL(R)"             "AL-VC1(R)"         "AL-D(R)"          
#>  [25] "AL-VA5(R)"         "RUB(R)"            "ATL(L)"           
#>  [28] "MB(L)"             "NO"                "LH(R)"            
#>  [31] "SLP(R)"            "AB(R)"             "MB(R)"            
#>  [34] "a'3(R)"            "b'2(R)"            "SNP(R)"           
#>  [37] "AL-DA3(L)"         "AL-VM7v(L)"        "AL-VC2(R)"        
#>  [40] "AL-DC2(L)"         "AL-DC4(L)"         "AL-VL2p(R)"       
#>  [43] "AL-VM3(R)"         "b2(R)"             "BU(R)"            
#>  [46] "ATL(R)"            "AL-DL3(R)"         "ME(R)"            
#>  [49] "INP"               "AL-DA1(R)"         "AL-DC1(R)"        
#>  [52] "LX(R)"             "FBl8"              "AL-VA6(R)"        
#>  [55] "EBr6"              "AL-VM1(R)"         "VLNP(R)"          
#>  [58] "PB(R6)"            "AL-VA4(R)"         "WED(R)"           
#>  [61] "gL(R)"             "gL(L)"             "MB(+ACA)(R)"      
#>  [64] "RUB(L)"            "PB(L3)"            "AL-DP1m(R)"       
#>  [67] "EPA(L)"            "AL-DM5(R)"         "AL-VA3(R)"        
#>  [70] "PENP"              "FLA(R)"            "EBr2r4"           
#>  [73] "FBl3"              "AME(R)"            "AL-DM2(R)"        
#>  [76] "EB"                "AL-DC4(R)"         "AL-DC3(R)"        
#>  [79] "AL-VC3l(R)"        "a1(R)"             "g3(R)"            
#>  [82] "PB"                "AL-DL5(L)"         "AL-DA2(R)"        
#>  [85] "CA(L)"             "lACA(R)"           "EBr1"             
#>  [88] "AL-VA6(L)"         "SNP(L)"            "AL-DM4(R)"        
#>  [91] "AL-VA7m(R)"        "AL-VP3(R)"         "mALT(R)"          
#>  [94] "AL-VP1l(R)"        "AVLP(R)"           "AL(R)"            
#>  [97] "AB(L)"             "EBr5"              "AL(L)"            
#> [100] "NO2(R)"            "GF(R)"             "AOTU(R)"          
#> [103] "AL-DM6(L)"         "LX(L)"             "AL-VP1m(R)"       
#> [106] "IB"                "GA(R)"             "PB(R9)"           
#> [109] "FBl2"              "AL-VA2(R)"         "FB"               
#> [112] "AL-V(R)"           "PB(L2)"            "g1(R)"            
#> [115] "AL-D(L)"           "CA(R)"             "AL-VM2(R)"        
#> [118] "EBr3am"            "dACA(R)"           "ROB(R)"           
#> [121] "AL-DA3(R)"         "AL-DL2v(R)"        "aL(R)"            
#> [124] "g4(R)"             "ICL(L)"            "FBl5"             
#> [127] "LOP(R)"            "SIP(R)"            "FBl4"             
#> [130] "b'L(L)"            "AL-VC4(R)"         "AL-VP4(R)"        
#> [133] "GC"                "g5(R)"             "AL-DL2d(R)"       
#> [136] "AL-VM7v(R)"        "AL-VL2a(R)"        "PB(R1)"           
#> [139] "FBl6"              "GOR(R)"            "AL-DL1(R)"        
#> [142] "AL-DC2(R)"         "AL-DM1(R)"         "AL-VM7d(R)"       
#> [145] "AL-DM5(L)"         "NO3(L)"            "b1(R)"            
#> [148] "IPS(R)"            "SMP(L)"            "AL-VA1v(R)"       
#> [151] "PRW"               "bL(R)"             "PB(R5)"           
#> [154] "SIP(L)"            "b'1(R)"            "PB(R8)"           
#> [157] "ICL(R)"            "aL(L)"             "AL-VP1d(R)"       
#> [160] "NO2(L)"            "PB(L1)"            "AL-DM3(L)"        
#> [163] "AL-VM7d(L)"        "AL-DM6(R)"         "hemibrain"        
#> [166] "CRE(L)"            "IB112"             "NO(L)"            
#> [169] "BU(L)"             "a'1(R)"            "VES(L)"           
#> [172] "AL-DA4m(R)"        "SAD(-AMMC)"        "SCL(R)"           
#> [175] "AL-DP1m(L)"        "CRE(-RUB)(L)"      "bL(L)"            
#> [178] "SAD"               "NO(R)"             "AOT(R)"           
#> [181] "vACA(R)"           "AL-VM5d(R)"        "AL-DM3(R)"        
#> [184] "AL-VC5(R)"         "AL-DA4m(L)"        "AL-DM2(L)"        
#> [187] "LAL(R)"            "AL-DL4(R)"         "SPS(R)"           
#> [190] "b'L(R)"            "AL-DC1(L)"         "CRE(-ROB,-RUB)(R)"
#> [193] "PB(L9)"            "FB-column3"        "a2(R)"            
#> [196] "LAL(-GA)(R)"       "SPS(L)"            "a'2(R)"           
#> [199] "a'L(R)"            "PB(R7)"            "CAN(R)"           
#> [202] "a'L(L)"            "AL-VM5v(R)"        "SMP(R)"           
#> [205] "LAL(L)"            "AL-DM1(L)"         "AL-DM4(L)"        
#> [208] "FBl1"              "CRE(R)"            "PB(R3)"           
#> [211] "AL-VA1d(R)"        "NO3(R)"            "PB(L8)"           
#> [214] "AL-VM4(R)"         "NO1(L)"            "PB(R2)"           
#> [217] "CX"                "PED(R)"            "PVLP(R)"          
#> [220] "mALT(L)"           "VMNP"              "AL-DP1l(R)"       
#> [223] "VES(R)"            "AL-VC3m(R)"        "NO1(R)"           
#> [226] "a3(R)"             "FBl7"              "g2(R)"            
#> [229] "EBr3d"             "EBr3pw"            "LO(R)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(L8)"            "ME(R)"             "SCL(L)"           
#>   [4] "PB(R3)"            "PB(L3)"            "WED(R)"           
#>   [7] "g4(R)"             "LH(R)"             "EPA(L)"           
#>  [10] "VLNP(R)"           "PB(R5)"            "GF(R)"            
#>  [13] "NO3(R)"            "FBl4"              "BU(R)"            
#>  [16] "GNG"               "bL(R)"             "b'1(R)"           
#>  [19] "IB"                "CRE(R)"            "dACA(R)"          
#>  [22] "EBr3d"             "mALT(L)"           "a1(R)"            
#>  [25] "a3(R)"             "LAL(R)"            "PB(R6)"           
#>  [28] "AL(L)"             "FBl6"              "EPA(R)"           
#>  [31] "MB(+ACA)(R)"       "FLA(R)"            "NO2(R)"           
#>  [34] "NO(L)"             "AOT(R)"            "a2(R)"            
#>  [37] "ROB(R)"            "AVLP(R)"           "a'1(R)"           
#>  [40] "b1(R)"             "EB"                "gL(L)"            
#>  [43] "PB(L9)"            "LX(L)"             "MB(R)"            
#>  [46] "hemibrain"         "SIP(R)"            "PB(L6)"           
#>  [49] "GOR(L)"            "b'L(R)"            "PB(R4)"           
#>  [52] "ICL(R)"            "AB(R)"             "CX"               
#>  [55] "SPS(L)"            "SNP(R)"            "FBl9"             
#>  [58] "VMNP"              "VES(R)"            "PB"               
#>  [61] "CRE(-ROB,-RUB)(R)" "GC"                "NO3(L)"           
#>  [64] "FBl8"              "SCL(R)"            "NO1(R)"           
#>  [67] "FBl1"              "IPS(R)"            "SMP(R)"           
#>  [70] "SIP(L)"            "lACA(R)"           "b2(R)"            
#>  [73] "ATL(R)"            "PB(L4)"            "OL(R)"            
#>  [76] "NO2(L)"            "NO1(L)"            "NO(R)"            
#>  [79] "SAD(-AMMC)"        "CA(L)"             "EBr6"             
#>  [82] "FBl7"              "POC"               "PRW"              
#>  [85] "mALT(R)"           "AMMC"              "RUB(R)"           
#>  [88] "PB(R7)"            "MB(L)"             "b'L(L)"           
#>  [91] "SNP(L)"            "LO(R)"             "CA(R)"            
#>  [94] "VES(L)"            "a'3(R)"            "AOTU(R)"          
#>  [97] "BU(L)"             "a'2(R)"            "PB(L5)"           
#> [100] "FBl5"              "AME(R)"            "ATL(L)"           
#> [103] "EBr1"              "NO"                "CRE(L)"           
#> [106] "PB(R1)"            "bL(L)"             "PENP"             
#> [109] "SAD"               "PB(R9)"            "g3(R)"            
#> [112] "PB(L7)"            "ICL(L)"            "SPS(R)"           
#> [115] "PB(L1)"            "LX(R)"             "RUB(L)"           
#> [118] "a'L(L)"            "LAL(L)"            "FB"               
#> [121] "PB(R2)"            "EBr2r4"            "FBl2"             
#> [124] "SLP(R)"            "LOP(R)"            "vACA(R)"          
#> [127] "b'2(R)"            "EBr5"              "GOR(R)"           
#> [130] "PLP(R)"            "a'L(R)"            "aL(L)"            
#> [133] "AL(R)"             "AB(L)"             "PED(R)"           
#> [136] "LAL(-GA)(R)"       "g2(R)"             "FBl3"             
#> [139] "PB(L2)"            "EBr3am"            "GA(R)"            
#> [142] "g1(R)"             "PB(R8)"            "EBr3pw"           
#> [145] "g5(R)"             "PVLP(R)"           "CAN(R)"           
#> [148] "INP"               "SMP(L)"            "aL(R)"            
#> [151] "gL(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "PB(L1)"            "VMNP"              "vACA(R)"          
#>   [4] "PB(R4)"            "PED(R)"            "PLP(R)"           
#>   [7] "PB(L2)"            "AL-DA2(L)"         "AL-DM6(L)"        
#>  [10] "NO3(L)"            "LX(L)"             "ROB(R)"           
#>  [13] "g3(R)"             "AOTU(R)"           "a'L(R)"           
#>  [16] "AL(L)"             "AB(R)"             "AL-DA3(R)"        
#>  [19] "AL-VA7m(R)"        "AL-VM3(R)"         "FBl8"             
#>  [22] "PB(R3)"            "AL-DA2(R)"         "AL-VM7v(L)"       
#>  [25] "VES(R)"            "g4(R)"             "ATL(R)"           
#>  [28] "SMP(R)"            "g2(R)"             "SCL(R)"           
#>  [31] "gL(L)"             "EPA(R)"            "AL-VA1d(R)"       
#>  [34] "EBr3d"             "EBr3pw"            "AL-DL2v(R)"       
#>  [37] "AL-DP1l(R)"        "gL(R)"             "bL(R)"            
#>  [40] "AMMC"              "PB(L9)"            "AME(R)"           
#>  [43] "AL-VL2a(R)"        "AL-DM2(L)"         "AL-DM4(L)"        
#>  [46] "CA(R)"             "AL-VA1v(R)"        "ME(R)"            
#>  [49] "AL-DA3(L)"         "AL-DC4(L)"         "AL-DL4(L)"        
#>  [52] "AL-VC3l(R)"        "PB(R5)"            "NO2(L)"           
#>  [55] "GF(R)"             "EBr5"              "AL-DC1(L)"        
#>  [58] "AL-VM5v(R)"        "MB(R)"             "AL-DA4l(R)"       
#>  [61] "CX"                "NO1(R)"            "aL(L)"            
#>  [64] "AL-V(R)"           "AL-VP4(R)"         "AL-DP1m(L)"       
#>  [67] "OL(R)"             "AL-VA7l(R)"        "RUB(R)"           
#>  [70] "CRE(R)"            "AL-DL2d(R)"        "mALT(L)"          
#>  [73] "GOR(L)"            "FBl6"              "GOR(R)"           
#>  [76] "AB(L)"             "AL-DM1(R)"         "a2(R)"            
#>  [79] "SAD"               "BU(L)"             "a'2(R)"           
#>  [82] "AL-VL1(R)"         "lACA(R)"           "mALT(R)"          
#>  [85] "SLP(R)"            "AL-VC2(R)"         "EBr3am"           
#>  [88] "a1(R)"             "b'L(R)"            "bL(L)"            
#>  [91] "a'L(L)"            "SCL(L)"            "CA(L)"            
#>  [94] "PB(R6)"            "AL-VM7d(L)"        "CRE(-RUB)(L)"     
#>  [97] "AL(R)"             "AL-DA4m(L)"        "ICL(R)"           
#> [100] "AL-VA4(R)"         "POC"               "aL(R)"            
#> [103] "BU(R)"             "AL-DL5(L)"         "INP"              
#> [106] "FBl2"              "AL-VM5d(R)"        "AL-VP3(R)"        
#> [109] "LH(R)"             "GA(R)"             "NO1(L)"           
#> [112] "LO(R)"             "PVLP(R)"           "AL-D(R)"          
#> [115] "LAL(L)"            "EB"                "LAL(R)"           
#> [118] "AL-DM2(R)"         "AL-DC2(R)"         "EBr1"             
#> [121] "AL-DM5(L)"         "PB(R9)"            "PB(L7)"           
#> [124] "AL-VA6(L)"         "ATL(L)"            "SAD(-AMMC)"       
#> [127] "NO(L)"             "AL-VM7v(R)"        "VLNP(R)"          
#> [130] "hemibrain"         "b2(R)"             "LX(R)"            
#> [133] "g5(R)"             "FBl9"              "AL-DL1(R)"        
#> [136] "IB"                "FB"                "EBr6"             
#> [139] "PB(L6)"            "VES(L)"            "NO2(R)"           
#> [142] "AL-VM2(R)"         "AL-VP2(R)"         "MB(+ACA)(R)"      
#> [145] "AL-DL4(R)"         "AL-DP1m(R)"        "AL-VC4(R)"        
#> [148] "b1(R)"             "AL-D(L)"           "AL-DM1(L)"        
#> [151] "AL-VA2(R)"         "AL-VP5(R)"         "AL-VP1l(R)"       
#> [154] "PB(L4)"            "AL-VC3m(R)"        "LAL(-GA)(R)"      
#> [157] "EBr2r4"            "FB-column3"        "WED(R)"           
#> [160] "RUB(L)"            "AL-VC5(R)"         "PB(L5)"           
#> [163] "AL-VM7d(R)"        "AL-VP1m(R)"        "AL-DC2(L)"        
#> [166] "CRE(L)"            "AL-DM3(R)"         "AL-DM5(R)"        
#> [169] "LOP(R)"            "CRE(-ROB,-RUB)(R)" "SMP(L)"           
#> [172] "dACA(R)"           "FBl5"              "AL-DL5(R)"        
#> [175] "AL-DM6(R)"         "SIP(R)"            "PB(R1)"           
#> [178] "FBl3"              "FBl4"              "SPS(R)"           
#> [181] "PB(R8)"            "AL-DA4m(R)"        "a'3(R)"           
#> [184] "NO(R)"             "GNG"               "b'1(R)"           
#> [187] "AL-VP1d(R)"        "FBl1"              "IPS(R)"           
#> [190] "AL-VA5(R)"         "AL-VM1(R)"         "EPA(L)"           
#> [193] "PB(R2)"            "AL-DC1(R)"         "AL-DC4(R)"        
#> [196] "ICL(L)"            "AL-DC3(R)"         "AL-DM4(R)"        
#> [199] "AL-VC1(R)"         "AL-DM3(L)"         "AL-VA6(R)"        
#> [202] "PENP"              "PB(R7)"            "MB(L)"            
#> [205] "SPS(L)"            "PRW"               "FBl7"             
#> [208] "b'L(L)"            "FLA(R)"            "PB(L8)"           
#> [211] "CAN(R)"            "AVLP(R)"           "SIP(L)"           
#> [214] "SNP(R)"            "AL-VM4(R)"         "AOT(R)"           
#> [217] "PB(L3)"            "g1(R)"             "a'1(R)"           
#> [220] "a3(R)"             "b'2(R)"            "GC"               
#> [223] "PB"                "AL-DA1(R)"         "NO3(R)"           
#> [226] "NO"                "SNP(L)"            "AL-DL3(R)"        
#> [229] "AL-VA3(R)"         "AL-VL2p(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "PB(R1)"            "bL(L)"             "PB(R2)"           
#>   [4] "PB(R6)"            "PB(L9)"            "AL-VL2p(R)"       
#>   [7] "BU(L)"             "AL-VM3(R)"         "RUB(R)"           
#>  [10] "AL-DP1l(R)"        "CA(L)"             "AL-DC4(L)"        
#>  [13] "VLNP(R)"           "AL-VP5(R)"         "CX"               
#>  [16] "GOR(L)"            "SIP(R)"            "SAD(-AMMC)"       
#>  [19] "AL-DC1(R)"         "AL-DM5(R)"         "AL-DM6(R)"        
#>  [22] "AL-VA7l(R)"        "AL-VC5(R)"         "AOTU(R)"          
#>  [25] "AL-VM7d(R)"        "SCL(R)"            "EBr6"             
#>  [28] "OL(R)"             "AL-DC2(R)"         "AL-VM5d(R)"       
#>  [31] "GF(R)"             "CRE(-RUB)(L)"      "AL-VM7d(L)"       
#>  [34] "AL-DA3(R)"         "PLP(R)"            "AL-VP1l(R)"       
#>  [37] "vACA(R)"           "NO(R)"             "mALT(L)"          
#>  [40] "IB"                "AL-DA2(L)"         "FBl5"             
#>  [43] "a'L(L)"            "b'1(R)"            "CRE(R)"           
#>  [46] "b1(R)"             "NO(L)"             "AL-DL2v(R)"       
#>  [49] "AL-VA7m(R)"        "MB(L)"             "a'L(R)"           
#>  [52] "AL-DA4m(L)"        "AL-DM4(L)"         "ICL(L)"           
#>  [55] "AL-DA4m(R)"        "g2(R)"             "EBr3pw"           
#>  [58] "EBr2r4"            "gL(L)"             "AL-DC4(R)"        
#>  [61] "AL-VM4(R)"         "AL-VP1m(R)"        "PB(L2)"           
#>  [64] "CRE(-ROB,-RUB)(R)" "g1(R)"             "ATL(R)"           
#>  [67] "EPA(L)"            "AL-VA4(R)"         "AL-VM1(R)"        
#>  [70] "FBl6"              "AL-DL3(R)"         "EBr3d"            
#>  [73] "AL-VM7v(L)"        "PB"                "AL-DM4(R)"        
#>  [76] "AL-DA1(R)"         "AL-DL4(L)"         "NO3(L)"           
#>  [79] "AB(R)"             "LX(R)"             "VES(L)"           
#>  [82] "AL-VA2(R)"         "NO3(R)"            "AL-DM5(L)"        
#>  [85] "MB(+ACA)(R)"       "FB-column3"        "AL-DP1m(R)"       
#>  [88] "AOT(R)"            "AMMC"              "PRW"              
#>  [91] "g5(R)"             "FBl1"              "GC"               
#>  [94] "AL-VM7v(R)"        "a1(R)"             "AL-VA5(R)"        
#>  [97] "FBl8"              "AME(R)"            "AL-VA3(R)"        
#> [100] "SNP(R)"            "LAL(-GA)(R)"       "PB(L8)"           
#> [103] "FBl4"              "a3(R)"             "SMP(R)"           
#> [106] "SAD"               "PB(L5)"            "FLA(R)"           
#> [109] "AL-DA4l(R)"        "PVLP(R)"           "PB(L6)"           
#> [112] "INP"               "LAL(R)"            "SCL(L)"           
#> [115] "SPS(L)"            "PB(L7)"            "a2(R)"            
#> [118] "PB(R8)"            "BU(R)"             "AL-DP1m(L)"       
#> [121] "RUB(L)"            "SLP(R)"            "AL-VA6(R)"        
#> [124] "AL-VC2(R)"         "PED(R)"            "NO2(L)"           
#> [127] "AL-VA1d(R)"        "IPS(R)"            "LX(L)"            
#> [130] "AL-VM2(R)"         "a'1(R)"            "AL-DL5(L)"        
#> [133] "AL-VA6(L)"         "AL-VP2(R)"         "b2(R)"            
#> [136] "AL-DL4(R)"         "CA(R)"             "ROB(R)"           
#> [139] "NO"                "AL-DC2(L)"         "AL-D(L)"          
#> [142] "AL-DM2(L)"         "AL-DM2(R)"         "AL-VL2a(R)"       
#> [145] "AL-DC1(L)"         "EBr5"              "AL-DM6(L)"        
#> [148] "CRE(L)"            "SIP(L)"            "ME(R)"            
#> [151] "gL(R)"             "LO(R)"             "AVLP(R)"          
#> [154] "WED(R)"            "PB(L1)"            "VMNP"             
#> [157] "PB(R4)"            "LAL(L)"            "bL(R)"            
#> [160] "PB(L4)"            "AL-DL2d(R)"        "AL-DM3(R)"        
#> [163] "AB(L)"             "VES(R)"            "AL-DM3(L)"        
#> [166] "CAN(R)"            "PB(L3)"            "AL-DM1(L)"        
#> [169] "AL-VL1(R)"         "AL-DA3(L)"         "AL-DL1(R)"        
#> [172] "AL-VA1v(R)"        "PB(R5)"            "AL-DA2(R)"        
#> [175] "FBl3"              "hemibrain"         "FBl9"             
#> [178] "AL-VC3l(R)"        "FB"                "aL(R)"            
#> [181] "b'L(R)"            "dACA(R)"           "g4(R)"            
#> [184] "AL-VC1(R)"         "AL-VM5v(R)"        "g3(R)"            
#> [187] "LOP(R)"            "GNG"               "a'2(R)"           
#> [190] "AL(L)"             "LH(R)"             "AL-VC3m(R)"       
#> [193] "AL-VP1d(R)"        "NO1(L)"            "GA(R)"            
#> [196] "EPA(R)"            "PENP"              "EBr1"             
#> [199] "NO1(R)"            "FBl7"              "b'L(L)"           
#> [202] "MB(R)"             "AL-VC4(R)"         "AL-VP4(R)"        
#> [205] "SPS(R)"            "SMP(L)"            "EBr3am"           
#> [208] "EB"                "SNP(L)"            "ICL(R)"           
#> [211] "aL(L)"             "PB(R9)"            "mALT(R)"          
#> [214] "PB(R3)"            "NO2(R)"            "GOR(R)"           
#> [217] "AL(R)"             "AL-DM1(R)"         "PB(R7)"           
#> [220] "AL-D(R)"           "b'2(R)"            "AL-DC3(R)"        
#> [223] "AL-V(R)"           "a'3(R)"            "FBl2"             
#> [226] "AL-DL5(R)"         "AL-VP3(R)"         "POC"              
#> [229] "lACA(R)"           "ATL(L)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method    path                                        
#> [1,] "CONNECT" "/api"                                      
#> [2,] "GET"     "/api/v:ver/npexplorer/completeness"        
#> [3,] "GET"     "/api/v:ver/skeletons/skeleton/:dataset/:id"
#> [4,] "POST"    "/api/roimeshes/mesh/:dataset/:roi"         
#> [5,] "POST"    "/api/npexplorer/roiconnectivity"           
#> [6,] "POST"    "/api/v:ver/raw/keyvalue/key/:instance/:key"
#>      name                                                                                           
#> [1,] "github.com/labstack/echo.glob..func1"                                                         
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                       
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIConnectivity-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }