Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "b2(R)"             "ME(R)"             "SIP(R)"           
#>   [4] "PB(R5)"            "SLP(R)"            "AL-DL2v(R)"       
#>   [7] "SPS(R)"            "SPS(L)"            "VMNP"             
#>  [10] "PB(L7)"            "EPA(L)"            "GOR(R)"           
#>  [13] "AL-VM3(R)"         "PB(L3)"            "EBr3d"            
#>  [16] "IB112"             "AL-DL2d(R)"        "mALT(L)"          
#>  [19] "ATL(L)"            "NO3(R)"            "CRE(-RUB)(L)"     
#>  [22] "b1(R)"             "AL-DA4m(R)"        "NO"               
#>  [25] "PVLP(R)"           "FB-column3"        "AL-VA7m(R)"       
#>  [28] "GC"                "FB"                "AL-DL4(L)"        
#>  [31] "AL-DC3(R)"         "PB(L2)"            "PB(R6)"           
#>  [34] "EBr1"              "NO(R)"             "AME(R)"           
#>  [37] "AL-VP3(R)"         "SNP(R)"            "AL-VA3(R)"        
#>  [40] "AL-DC2(L)"         "AL-VP2(R)"         "CRE(-ROB,-RUB)(R)"
#>  [43] "LAL(-GA)(R)"       "NO1(R)"            "AL-VM7v(L)"       
#>  [46] "FBl8"              "AL-DL5(R)"         "AL-VP1l(R)"       
#>  [49] "a1(R)"             "a'L(R)"            "PLP(R)"           
#>  [52] "FBl9"              "AL-VA7l(R)"        "AL-VC5(R)"        
#>  [55] "AL-VL2a(R)"        "AL-VM5v(R)"        "b'2(R)"           
#>  [58] "MB(+ACA)(R)"       "FBl3"              "MB(L)"            
#>  [61] "AL-DM4(L)"         "AL-VA6(L)"         "AL-DM6(R)"        
#>  [64] "AL-DP1l(R)"        "AL-VA2(R)"         "g3(R)"            
#>  [67] "GA(R)"             "gL(L)"             "EBr3am"           
#>  [70] "mALT(R)"           "AL-DM6(L)"         "g5(R)"            
#>  [73] "IPS(R)"            "GOR(L)"            "VES(L)"           
#>  [76] "AL-VM5d(R)"        "dACA(R)"           "AL-DM5(L)"        
#>  [79] "LX(L)"             "AL-D(R)"           "AL-VA6(R)"        
#>  [82] "LAL(L)"            "ATL(R)"            "AL-DC4(R)"        
#>  [85] "AL-DP1m(R)"        "AL-VP1d(R)"        "POC"              
#>  [88] "FLA(R)"            "LX(R)"             "SNP(L)"           
#>  [91] "AL-DC2(R)"         "vACA(R)"           "b'L(R)"           
#>  [94] "FBl2"              "AL-DM4(R)"         "NO1(L)"           
#>  [97] "AL-DM2(L)"         "FBl4"              "AMMC"             
#> [100] "a'2(R)"            "EB"                "b'L(L)"           
#> [103] "NO(L)"             "AL-DL4(R)"         "VES(R)"           
#> [106] "AL-VM4(R)"         "SMP(L)"            "EBr2r4"           
#> [109] "AL-DA3(L)"         "AL-DA4m(L)"        "LH(R)"            
#> [112] "AL-VL2p(R)"        "bL(R)"             "AB(L)"            
#> [115] "PB(L5)"            "AL-VP1m(R)"        "AOTU(R)"          
#> [118] "PB(L6)"            "aL(L)"             "AL-VA1d(R)"       
#> [121] "AL-VA1v(R)"        "AL-VC4(R)"         "AL-VM7d(R)"       
#> [124] "LOP(R)"            "gL(R)"             "AL-VM7d(L)"       
#> [127] "INP"               "a2(R)"             "FBl7"             
#> [130] "AL(L)"             "AB(R)"             "PB(R7)"           
#> [133] "PB(R9)"            "ROB(R)"            "FBl1"             
#> [136] "PB(L4)"            "AL-DA1(R)"         "a'L(L)"           
#> [139] "MB(R)"             "AL-V(R)"           "PENP"             
#> [142] "AL-DP1m(L)"        "PB(R1)"            "OL(R)"            
#> [145] "IB"                "AL-DC4(L)"         "NO3(L)"           
#> [148] "RUB(R)"            "lACA(R)"           "PB(L8)"           
#> [151] "BU(R)"             "AL-DA2(L)"         "AL-VA5(R)"        
#> [154] "SMP(R)"            "hemibrain"         "EBr6"             
#> [157] "CRE(L)"            "AL-VM1(R)"         "NO2(L)"           
#> [160] "g1(R)"             "EBr5"              "AL-DM3(L)"        
#> [163] "AL-VC1(R)"         "bL(L)"             "AL-DM1(R)"        
#> [166] "GF(R)"             "SAD"               "CRE(R)"           
#> [169] "CA(R)"             "CX"                "FBl6"             
#> [172] "CAN(R)"            "AL(R)"             "SCL(L)"           
#> [175] "PB(R8)"            "LO(R)"             "ICL(R)"           
#> [178] "AL-DM5(R)"         "AL-VC3m(R)"        "AL-VM2(R)"        
#> [181] "g2(R)"             "BU(L)"             "AL-DM2(R)"        
#> [184] "aL(R)"             "PB(L9)"            "PB(R2)"           
#> [187] "AOT(R)"            "WED(R)"            "PB(L1)"           
#> [190] "AL-DL5(L)"         "AL-DA3(R)"         "AL-D(L)"          
#> [193] "ICL(L)"            "NO2(R)"            "LAL(R)"           
#> [196] "PED(R)"            "SAD(-AMMC)"        "PRW"              
#> [199] "AL-DA2(R)"         "AL-VC2(R)"         "AL-VP4(R)"        
#> [202] "PB"                "AL-DC1(L)"         "AL-DM1(L)"        
#> [205] "PB(R3)"            "RUB(L)"            "VLNP(R)"          
#> [208] "b'1(R)"            "AL-DA4l(R)"        "AL-VA4(R)"        
#> [211] "AL-VM7v(R)"        "AL-VP5(R)"         "a'3(R)"           
#> [214] "SCL(R)"            "a'1(R)"            "AL-DC1(R)"        
#> [217] "AL-VC3l(R)"        "AL-VL1(R)"         "SIP(L)"           
#> [220] "GNG"               "AL-DL1(R)"         "AL-DM3(R)"        
#> [223] "AVLP(R)"           "PB(R4)"            "g4(R)"            
#> [226] "EPA(R)"            "EBr3pw"            "FBl5"             
#> [229] "a3(R)"             "AL-DL3(R)"         "CA(L)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(L1)"            "gL(L)"             "g5(R)"            
#>   [4] "AL(R)"             "b'1(R)"            "GC"               
#>   [7] "MB(L)"             "SCL(R)"            "hemibrain"        
#>  [10] "EPA(L)"            "ATL(R)"            "INP"              
#>  [13] "RUB(L)"            "a3(R)"             "vACA(R)"          
#>  [16] "FB"                "PB"                "gL(R)"            
#>  [19] "PVLP(R)"           "FBl3"              "CAN(R)"           
#>  [22] "CRE(-ROB,-RUB)(R)" "SPS(R)"            "LX(R)"            
#>  [25] "EBr2r4"            "LO(R)"             "a'3(R)"           
#>  [28] "NO2(L)"            "PB(R7)"            "FLA(R)"           
#>  [31] "PB(R8)"            "a'2(R)"            "SIP(R)"           
#>  [34] "VES(R)"            "IB"                "b'2(R)"           
#>  [37] "a2(R)"             "b'L(R)"            "POC"              
#>  [40] "AOT(R)"            "SMP(R)"            "PB(L2)"           
#>  [43] "PB(L7)"            "FBl2"              "SLP(R)"           
#>  [46] "EBr1"              "EBr5"              "PB(L5)"           
#>  [49] "BU(L)"             "FBl4"              "NO3(R)"           
#>  [52] "EPA(R)"            "AL(L)"             "g3(R)"            
#>  [55] "PB(R4)"            "PB(L6)"            "b'L(L)"           
#>  [58] "EBr3am"            "SAD"               "NO1(L)"           
#>  [61] "dACA(R)"           "g1(R)"             "g4(R)"            
#>  [64] "MB(+ACA)(R)"       "aL(L)"             "PB(L4)"           
#>  [67] "PB(L3)"            "lACA(R)"           "WED(R)"           
#>  [70] "CA(L)"             "CRE(R)"            "SPS(L)"           
#>  [73] "NO2(R)"            "NO"                "FBl8"             
#>  [76] "LAL(R)"            "VES(L)"            "GF(R)"            
#>  [79] "PLP(R)"            "VMNP"              "SCL(L)"           
#>  [82] "a1(R)"             "NO1(R)"            "SNP(R)"           
#>  [85] "a'L(L)"            "PB(L8)"            "FBl9"             
#>  [88] "AB(L)"             "PENP"              "VLNP(R)"          
#>  [91] "AMMC"              "FBl5"              "PB(R3)"           
#>  [94] "GNG"               "GOR(R)"            "mALT(L)"          
#>  [97] "LAL(-GA)(R)"       "PB(R1)"            "PB(L9)"           
#> [100] "MB(R)"             "b2(R)"             "SAD(-AMMC)"       
#> [103] "b1(R)"             "SNP(L)"            "AB(R)"            
#> [106] "ATL(L)"            "PRW"               "g2(R)"            
#> [109] "EB"                "NO(L)"             "CX"               
#> [112] "BU(R)"             "bL(L)"             "FBl7"             
#> [115] "EBr3pw"            "PB(R2)"            "LX(L)"            
#> [118] "LOP(R)"            "SMP(L)"            "AOTU(R)"          
#> [121] "ME(R)"             "PB(R5)"            "EBr6"             
#> [124] "AVLP(R)"           "ICL(R)"            "NO3(L)"           
#> [127] "ICL(L)"            "CA(R)"             "a'1(R)"           
#> [130] "FBl6"              "IPS(R)"            "GA(R)"            
#> [133] "ROB(R)"            "NO(R)"             "LAL(L)"           
#> [136] "GOR(L)"            "AME(R)"            "a'L(R)"           
#> [139] "CRE(L)"            "PB(R6)"            "PED(R)"           
#> [142] "LH(R)"             "FBl1"              "bL(R)"            
#> [145] "RUB(R)"            "mALT(R)"           "aL(R)"            
#> [148] "PB(R9)"            "SIP(L)"            "EBr3d"            
#> [151] "OL(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "NO"                "EPA(L)"            "EB"               
#>   [4] "AL(L)"             "SLP(R)"            "AL-VL2a(R)"       
#>   [7] "g3(R)"             "dACA(R)"           "INP"              
#>  [10] "SMP(R)"            "aL(R)"             "FBl5"             
#>  [13] "ATL(R)"            "AL-DA4m(R)"        "hemibrain"        
#>  [16] "AL-VA1v(R)"        "SAD"               "CRE(L)"           
#>  [19] "AL-DA3(R)"         "PLP(R)"            "SAD(-AMMC)"       
#>  [22] "LH(R)"             "IB"                "PB(L2)"           
#>  [25] "mALT(L)"           "NO1(R)"            "FBl8"             
#>  [28] "SMP(L)"            "AL-VA6(L)"         "AL-DL3(R)"        
#>  [31] "AL-VC2(R)"         "PENP"              "MB(L)"            
#>  [34] "MB(+ACA)(R)"       "LAL(L)"            "PB"               
#>  [37] "AL-VC3m(R)"        "NO1(L)"            "AL-DL5(L)"        
#>  [40] "AVLP(R)"           "PB(L4)"            "OL(R)"            
#>  [43] "AL-VM5v(R)"        "mALT(R)"           "FBl7"             
#>  [46] "bL(R)"             "AL-D(R)"           "GC"               
#>  [49] "AL-DM3(R)"         "AL-DM4(R)"         "PB(L8)"           
#>  [52] "FB"                "AL-DM5(L)"         "PB(R2)"           
#>  [55] "AL-VL2p(R)"        "PB(R6)"            "FBl2"             
#>  [58] "AL-VA3(R)"         "NO2(L)"            "BU(L)"            
#>  [61] "EBr2r4"            "NO3(L)"            "AL-VM3(R)"        
#>  [64] "AB(L)"             "NO3(R)"            "g2(R)"            
#>  [67] "SIP(L)"            "AMMC"              "a2(R)"            
#>  [70] "AL-DC3(R)"         "AL-DL2v(R)"        "AL-DM6(R)"        
#>  [73] "a'1(R)"            "bL(L)"             "LX(L)"            
#>  [76] "AL-DC2(R)"         "AL-VA4(R)"         "a'3(R)"           
#>  [79] "GF(R)"             "AL-DL5(R)"         "ICL(R)"           
#>  [82] "VES(R)"            "AL-VC3l(R)"        "AL-VC4(R)"        
#>  [85] "a'2(R)"            "EBr1"              "AME(R)"           
#>  [88] "aL(L)"             "AL-VP2(R)"         "lACA(R)"          
#>  [91] "EBr5"              "EBr3d"             "SNP(R)"           
#>  [94] "a'L(R)"            "AL-DL4(L)"         "FBl3"             
#>  [97] "FBl9"              "SCL(R)"            "AL-DA4l(R)"       
#> [100] "PB(R1)"            "LAL(R)"            "AL-VM1(R)"        
#> [103] "AL-VP4(R)"         "ROB(R)"            "AL-VC5(R)"        
#> [106] "LX(R)"             "AL-DP1m(L)"        "SIP(R)"           
#> [109] "AB(R)"             "AL-VM5d(R)"        "LOP(R)"           
#> [112] "g1(R)"             "ICL(L)"            "AL-DM5(R)"        
#> [115] "AL-DP1m(R)"        "AL-VA7m(R)"        "AL-VM7v(R)"       
#> [118] "PB(R8)"            "AL-DA3(L)"         "PB(R9)"           
#> [121] "AL-VP3(R)"         "BU(R)"             "AL-DM1(L)"        
#> [124] "CA(R)"             "SCL(L)"            "AL-DA2(L)"        
#> [127] "WED(R)"            "RUB(L)"            "SPS(R)"           
#> [130] "gL(L)"             "FBl4"              "PB(R3)"           
#> [133] "AL-DA1(R)"         "EBr3am"            "EBr6"             
#> [136] "gL(R)"             "NO2(R)"            "PB(R4)"           
#> [139] "AL-VA7l(R)"        "PB(R5)"            "CX"               
#> [142] "b'2(R)"            "AL-DC4(L)"         "AL-DM1(R)"        
#> [145] "PB(L5)"            "AL-DA4m(L)"        "FBl1"             
#> [148] "AL-DP1l(R)"        "AOTU(R)"           "AL(R)"            
#> [151] "AL-VM7v(L)"        "AL-DL1(R)"         "AL-DM2(L)"        
#> [154] "NO(R)"             "AL-VA2(R)"         "AL-V(R)"          
#> [157] "ATL(L)"            "GA(R)"             "AL-DM4(L)"        
#> [160] "PB(L9)"            "AL-DC1(R)"         "POC"              
#> [163] "vACA(R)"           "SPS(L)"            "FBl6"             
#> [166] "a'L(L)"            "AL-DL4(R)"         "AL-DM2(R)"        
#> [169] "PRW"               "EPA(R)"            "AL-DC2(L)"        
#> [172] "PVLP(R)"           "PB(R7)"            "AL-VP1l(R)"       
#> [175] "VES(L)"            "AL-DL2d(R)"        "AOT(R)"           
#> [178] "CA(L)"             "GOR(R)"            "AL-DC1(L)"        
#> [181] "AL-DM6(L)"         "a3(R)"             "AL-VA1d(R)"       
#> [184] "AL-VM4(R)"         "AL-VP1m(R)"        "PED(R)"           
#> [187] "AL-VM2(R)"         "GOR(L)"            "AL-VC1(R)"        
#> [190] "AL-VM7d(L)"        "IPS(R)"            "AL-VL1(R)"        
#> [193] "PB(L1)"            "AL-VP1d(R)"        "b'L(R)"           
#> [196] "FLA(R)"            "AL-D(L)"           "LO(R)"            
#> [199] "b'L(L)"            "NO(L)"             "b'1(R)"           
#> [202] "LAL(-GA)(R)"       "AL-DM3(L)"         "CRE(-RUB)(L)"     
#> [205] "EBr3pw"            "b1(R)"             "b2(R)"            
#> [208] "g4(R)"             "PB(L6)"            "SNP(L)"           
#> [211] "FB-column3"        "MB(R)"             "VLNP(R)"          
#> [214] "VMNP"              "RUB(R)"            "CAN(R)"           
#> [217] "g5(R)"             "GNG"               "PB(L7)"           
#> [220] "AL-VA6(R)"         "CRE(R)"            "AL-DC4(R)"        
#> [223] "AL-VA5(R)"         "AL-VM7d(R)"        "PB(L3)"           
#> [226] "a1(R)"             "ME(R)"             "AL-VP5(R)"        
#> [229] "CRE(-ROB,-RUB)(R)" "AL-DA2(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "NO3(R)"            "VES(R)"            "PB(R2)"           
#>   [4] "AL-DP1m(R)"        "AL-VM7d(R)"        "SCL(L)"           
#>   [7] "EPA(L)"            "a'3(R)"            "GC"               
#>  [10] "PB(L5)"            "GOR(L)"            "AB(L)"            
#>  [13] "SCL(R)"            "NO3(L)"            "b2(R)"            
#>  [16] "FBl6"              "CA(R)"             "AL-VM4(R)"        
#>  [19] "b'1(R)"            "LO(R)"             "PB(L4)"           
#>  [22] "AL-DC3(R)"         "AL-VM7v(L)"        "AL-VA5(R)"        
#>  [25] "AL-VM7v(R)"        "SMP(R)"            "b'2(R)"           
#>  [28] "LAL(-GA)(R)"       "PB"                "LX(R)"            
#>  [31] "VES(L)"            "AL-VA2(R)"         "PENP"             
#>  [34] "a2(R)"             "MB(L)"             "AL-VM7d(L)"       
#>  [37] "SAD"               "EBr2r4"            "AL-DP1m(L)"       
#>  [40] "AL-VM2(R)"         "AL-DC4(L)"         "ICL(L)"           
#>  [43] "POC"               "CRE(L)"            "AL-DL1(R)"        
#>  [46] "AL-VA6(R)"         "AL-VP2(R)"         "IB"               
#>  [49] "vACA(R)"           "gL(R)"             "AL-DA3(R)"        
#>  [52] "AL-VP3(R)"         "AL-VP1d(R)"        "LH(R)"            
#>  [55] "AL-DM1(R)"         "PB(R6)"            "hemibrain"        
#>  [58] "gL(L)"             "AL-DL4(L)"         "AL-VA6(L)"        
#>  [61] "PB(L7)"            "OL(R)"             "AL-DA1(R)"        
#>  [64] "g1(R)"             "PVLP(R)"           "AL-DM3(R)"        
#>  [67] "a3(R)"             "AL-DA4l(R)"        "GF(R)"            
#>  [70] "EBr6"              "SNP(R)"            "SNP(L)"           
#>  [73] "LOP(R)"            "VMNP"              "EPA(R)"           
#>  [76] "AL-DM2(L)"         "CRE(-RUB)(L)"      "PB(R7)"           
#>  [79] "NO1(R)"            "FBl1"              "ME(R)"            
#>  [82] "AL-DC4(R)"         "AL-VC1(R)"         "VLNP(R)"          
#>  [85] "PB(L2)"            "g5(R)"             "LAL(R)"           
#>  [88] "AL(R)"             "AL-DL2v(R)"        "AL-VC5(R)"        
#>  [91] "CRE(R)"            "PB(L8)"            "AL(L)"            
#>  [94] "FBl7"              "AL-DC1(R)"         "AL-DL3(R)"        
#>  [97] "PED(R)"            "EBr3d"             "AL-DA4m(L)"       
#> [100] "FBl5"              "SIP(L)"            "dACA(R)"          
#> [103] "AB(R)"             "mALT(L)"           "PB(L1)"           
#> [106] "FBl8"              "AL-DP1l(R)"        "AL-VA4(R)"        
#> [109] "AL-VP4(R)"         "CRE(-ROB,-RUB)(R)" "EBr3pw"           
#> [112] "FBl4"              "FBl9"              "AME(R)"           
#> [115] "AL-D(R)"           "AL-V(R)"           "AL-DA4m(R)"       
#> [118] "AL-VM3(R)"         "WED(R)"            "AL-DM5(L)"        
#> [121] "AL-DM6(L)"         "PB(R3)"            "AL-DA2(R)"        
#> [124] "SIP(R)"            "NO(L)"             "SLP(R)"           
#> [127] "AL-DM5(R)"         "AL-VM1(R)"         "EBr3am"           
#> [130] "PRW"               "AL-DA3(L)"         "AL-DC2(R)"        
#> [133] "AMMC"              "PB(L9)"            "CAN(R)"           
#> [136] "a'2(R)"            "AL-DL2d(R)"        "AL-VA1d(R)"       
#> [139] "AL-VP1l(R)"        "PB(L6)"            "NO2(R)"           
#> [142] "RUB(L)"            "ATL(L)"            "PB(L3)"           
#> [145] "GA(R)"             "AL-DL5(L)"         "FBl2"             
#> [148] "AL-DC2(L)"         "LX(L)"             "LAL(L)"           
#> [151] "NO1(L)"            "b'L(R)"            "NO(R)"            
#> [154] "aL(L)"             "AL-D(L)"           "b1(R)"            
#> [157] "mALT(R)"           "PB(R1)"            "SAD(-AMMC)"       
#> [160] "PB(R5)"            "NO2(L)"            "g3(R)"            
#> [163] "AL-DM1(L)"         "FBl3"              "ICL(R)"           
#> [166] "AL-VA7m(R)"        "AL-VL1(R)"         "EBr1"             
#> [169] "SPS(R)"            "AL-VA1v(R)"        "g2(R)"            
#> [172] "AL-DM3(L)"         "AL-DM4(L)"         "ROB(R)"           
#> [175] "AL-DL4(R)"         "AL-VC3m(R)"        "FB"               
#> [178] "EB"                "AL-VA3(R)"         "FB-column3"       
#> [181] "bL(R)"             "AL-VC4(R)"         "AOTU(R)"          
#> [184] "BU(L)"             "PB(R9)"            "aL(R)"            
#> [187] "b'L(L)"            "AL-DM4(R)"         "AL-VM5d(R)"       
#> [190] "RUB(R)"            "PB(R8)"            "bL(L)"            
#> [193] "AL-DM6(R)"         "AOT(R)"            "AL-DA2(L)"        
#> [196] "AL-DC1(L)"         "PB(R4)"            "MB(R)"            
#> [199] "AL-VP5(R)"         "GOR(R)"            "AL-VC2(R)"        
#> [202] "AL-VL2a(R)"        "a1(R)"             "AL-DL5(R)"        
#> [205] "CX"                "EBr5"              "FLA(R)"           
#> [208] "BU(R)"             "a'L(L)"            "AL-VM5v(R)"       
#> [211] "MB(+ACA)(R)"       "AVLP(R)"           "INP"              
#> [214] "AL-DM2(R)"         "AL-VP1m(R)"        "NO"               
#> [217] "ATL(R)"            "lACA(R)"           "SPS(L)"           
#> [220] "PLP(R)"            "a'L(R)"            "CA(L)"            
#> [223] "g4(R)"             "a'1(R)"            "IPS(R)"           
#> [226] "GNG"               "AL-VC3l(R)"        "AL-VL2p(R)"       
#> [229] "SMP(L)"            "AL-VA7l(R)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                        
#> [1,] "GET"  "/api/v:ver/dbmeta/instances"               
#> [2,] "GET"  "/api/v:ver/skeletons/skeleton/:dataset/:id"
#> [3,] "GET"  "/api/serverinfo"                           
#> [4,] "GET"  "/api/dbmeta/database"                      
#> [5,] "GET"  "/api/v:ver/npexplorer/completeness"        
#> [6,] "GET"  "/login"                                    
#>      name                                                                            
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"            
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"            
#> [3,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"        
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatabase-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"            
#> [6,] "github.com/janelia-flyem/echo-secure.loginHandler"                             

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }