Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DC3(R)"         "POC"               "RUB(R)"           
#>   [4] "LAL(L)"            "FB"                "VMNP"             
#>   [7] "AL-DA3(L)"         "FBl9"              "a'L(R)"           
#>  [10] "aL(L)"             "AL-DM5(R)"         "a'2(R)"           
#>  [13] "a'L(L)"            "AL-VP5(R)"         "SAD(-AMMC)"       
#>  [16] "AL-VA6(L)"         "SIP(R)"            "AL-DA4m(R)"       
#>  [19] "EB"                "PB(R1)"            "PENP"             
#>  [22] "FBl6"              "AL-DP1m(R)"        "AL-VA7l(R)"       
#>  [25] "IB"                "g2(R)"             "EPA(R)"           
#>  [28] "AL-VA1v(R)"        "AL-VM5d(R)"        "bL(L)"            
#>  [31] "AL-DA1(R)"         "AL-DA2(R)"         "AL-VM7v(L)"       
#>  [34] "AL-DC2(R)"         "AL-VA3(R)"         "AL-VC2(R)"        
#>  [37] "PB(L5)"            "RUB(L)"            "g1(R)"            
#>  [40] "SNP(R)"            "MB(+ACA)(R)"       "AL-DM1(R)"        
#>  [43] "CX"                "GF(R)"             "PLP(R)"           
#>  [46] "FBl1"              "AL-DL4(R)"         "AL-DC2(L)"        
#>  [49] "PB(R9)"            "VES(R)"            "AL-VM5v(R)"       
#>  [52] "FBl8"              "VES(L)"            "AL-DA4l(R)"       
#>  [55] "vACA(R)"           "SPS(L)"            "IB112"            
#>  [58] "AL-D(R)"           "AL-VA5(R)"         "NO2(L)"           
#>  [61] "FB-column3"        "FBl3"              "IPS(R)"           
#>  [64] "AL-DL3(R)"         "AL-VM7v(R)"        "NO1(L)"           
#>  [67] "SCL(R)"            "FLA(R)"            "FBl5"             
#>  [70] "AL-VC5(R)"         "LX(R)"             "AL-DM6(L)"        
#>  [73] "AL-DM3(R)"         "AL-DA2(L)"         "AL-VA7m(R)"       
#>  [76] "AL-DC4(R)"         "AL-VM3(R)"         "AL-VP3(R)"        
#>  [79] "mALT(L)"           "NO3(L)"            "AVLP(R)"          
#>  [82] "EPA(L)"            "PB(L8)"            "PB"               
#>  [85] "SLP(R)"            "AL-DP1l(R)"        "SCL(L)"           
#>  [88] "AL(R)"             "AL-DC1(R)"         "AL-VM2(R)"        
#>  [91] "g3(R)"             "MB(L)"             "AL-DA4m(L)"       
#>  [94] "PB(R3)"            "CAN(R)"            "AL-DM1(L)"        
#>  [97] "AL-VM7d(R)"        "CRE(-ROB,-RUB)(R)" "AL-DM2(L)"        
#> [100] "LH(R)"             "AL-VL2p(R)"        "AB(L)"            
#> [103] "BU(R)"             "AL(L)"             "ROB(R)"           
#> [106] "GOR(R)"            "AL-DM2(R)"         "PB(L3)"           
#> [109] "GOR(L)"            "AL-DL4(L)"         "PB(L9)"           
#> [112] "EBr2r4"            "NO"                "g5(R)"            
#> [115] "AL-VP2(R)"         "PB(R6)"            "PB(R4)"           
#> [118] "AL-VC3m(R)"        "EBr3d"             "AME(R)"           
#> [121] "AL-VA1d(R)"        "AL-VM7d(L)"        "AL-VC3l(R)"       
#> [124] "AL-VP1l(R)"        "GNG"               "SIP(L)"           
#> [127] "SPS(R)"            "NO1(R)"            "NO(R)"            
#> [130] "gL(R)"             "lACA(R)"           "a1(R)"            
#> [133] "PED(R)"            "AL-DA3(R)"         "LAL(-GA)(R)"      
#> [136] "LO(R)"             "EBr3am"            "FBl7"             
#> [139] "AL-DM6(R)"         "hemibrain"         "PB(R7)"           
#> [142] "b2(R)"             "PB(L1)"            "AL-DM5(L)"        
#> [145] "AL-DC4(L)"         "CRE(L)"            "PB(R2)"           
#> [148] "AL-VA6(R)"         "AL-VL2a(R)"        "AL-DP1m(L)"       
#> [151] "LOP(R)"            "GC"                "WED(R)"           
#> [154] "PVLP(R)"           "FBl4"              "MB(R)"            
#> [157] "a2(R)"             "mALT(R)"           "CRE(R)"           
#> [160] "PB(L6)"            "b'1(R)"            "dACA(R)"          
#> [163] "AB(R)"             "SAD"               "NO3(R)"           
#> [166] "NO2(R)"            "LAL(R)"            "CA(L)"            
#> [169] "VLNP(R)"           "EBr5"              "ME(R)"            
#> [172] "LX(L)"             "SMP(R)"            "a'3(R)"           
#> [175] "a3(R)"             "AOT(R)"            "PB(R8)"           
#> [178] "AL-DM4(L)"         "CRE(-RUB)(L)"      "ICL(L)"           
#> [181] "INP"               "AL-VA2(R)"         "AL-VL1(R)"        
#> [184] "AL-V(R)"           "b'2(R)"            "AL-DM3(L)"        
#> [187] "PB(L7)"            "SNP(L)"            "NO(L)"            
#> [190] "b'L(R)"            "EBr6"              "b1(R)"            
#> [193] "b'L(L)"            "PB(L2)"            "PRW"              
#> [196] "ATL(R)"            "AL-DL2d(R)"        "gL(L)"            
#> [199] "EBr3pw"            "CA(R)"             "AL-DL1(R)"        
#> [202] "AL-DL5(R)"         "g4(R)"             "AL-DL2v(R)"       
#> [205] "AL-VP4(R)"         "AL-VP1m(R)"        "AL-VP1d(R)"       
#> [208] "AL-VC4(R)"         "OL(R)"             "AL-D(L)"          
#> [211] "a'1(R)"            "ICL(R)"            "PB(L4)"           
#> [214] "PB(R5)"            "EBr1"              "AL-DL5(L)"        
#> [217] "FBl2"              "AL-DM4(R)"         "AOTU(R)"          
#> [220] "GA(R)"             "AL-DC1(L)"         "AL-VM4(R)"        
#> [223] "AMMC"              "bL(R)"             "AL-VC1(R)"        
#> [226] "ATL(L)"            "SMP(L)"            "aL(R)"            
#> [229] "BU(L)"             "AL-VA4(R)"         "AL-VM1(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "b1(R)"             "INP"               "a'L(L)"           
#>   [4] "SPS(L)"            "CRE(L)"            "SAD(-AMMC)"       
#>   [7] "PB(R3)"            "SAD"               "SNP(R)"           
#>  [10] "LOP(R)"            "ROB(R)"            "PB(L6)"           
#>  [13] "b'2(R)"            "aL(R)"             "MB(L)"            
#>  [16] "PB"                "PB(L5)"            "GOR(L)"           
#>  [19] "bL(L)"             "FBl8"              "a3(R)"            
#>  [22] "SIP(L)"            "FLA(R)"            "ICL(R)"           
#>  [25] "NO"                "MB(+ACA)(R)"       "PB(L4)"           
#>  [28] "RUB(R)"            "SIP(R)"            "PB(L9)"           
#>  [31] "NO3(L)"            "mALT(R)"           "EBr2r4"           
#>  [34] "AL(L)"             "RUB(L)"            "FBl5"             
#>  [37] "AL(R)"             "AOTU(R)"           "CRE(R)"           
#>  [40] "a'L(R)"            "ICL(L)"            "SCL(L)"           
#>  [43] "CRE(-ROB,-RUB)(R)" "GF(R)"             "NO3(R)"           
#>  [46] "PB(L8)"            "PB(L7)"            "AVLP(R)"          
#>  [49] "PB(R4)"            "SLP(R)"            "a1(R)"            
#>  [52] "FB"                "NO1(R)"            "g5(R)"            
#>  [55] "FBl2"              "AME(R)"            "PB(R5)"           
#>  [58] "g3(R)"             "a2(R)"             "gL(R)"            
#>  [61] "FBl3"              "g1(R)"             "EBr5"             
#>  [64] "mALT(L)"           "CA(L)"             "AOT(R)"           
#>  [67] "AB(L)"             "IB"                "ATL(L)"           
#>  [70] "LAL(L)"            "WED(R)"            "bL(R)"            
#>  [73] "LAL(R)"            "BU(R)"             "PB(R9)"           
#>  [76] "SNP(L)"            "VES(R)"            "PRW"              
#>  [79] "GC"                "NO1(L)"            "a'2(R)"           
#>  [82] "CAN(R)"            "aL(L)"             "GOR(R)"           
#>  [85] "b'1(R)"            "CX"                "EBr6"             
#>  [88] "g4(R)"             "a'1(R)"            "PB(R2)"           
#>  [91] "EBr3pw"            "IPS(R)"            "GNG"              
#>  [94] "ME(R)"             "OL(R)"             "SMP(L)"           
#>  [97] "SPS(R)"            "g2(R)"             "BU(L)"            
#> [100] "b'L(R)"            "EBr1"              "b'L(L)"           
#> [103] "MB(R)"             "LX(R)"             "LO(R)"            
#> [106] "LX(L)"             "b2(R)"             "PB(L1)"           
#> [109] "PLP(R)"            "FBl9"              "GA(R)"            
#> [112] "hemibrain"         "PB(R7)"            "PED(R)"           
#> [115] "ATL(R)"            "NO(L)"             "PB(R6)"           
#> [118] "PB(L2)"            "SCL(R)"            "NO(R)"            
#> [121] "LH(R)"             "NO2(R)"            "VLNP(R)"          
#> [124] "EBr3am"            "dACA(R)"           "EPA(R)"           
#> [127] "PVLP(R)"           "FBl6"              "PENP"             
#> [130] "PB(L3)"            "VMNP"              "NO2(L)"           
#> [133] "SMP(R)"            "LAL(-GA)(R)"       "gL(L)"            
#> [136] "AB(R)"             "EPA(L)"            "EBr3d"            
#> [139] "FBl7"              "VES(L)"            "a'3(R)"           
#> [142] "PB(R8)"            "FBl4"              "AMMC"             
#> [145] "CA(R)"             "POC"               "vACA(R)"          
#> [148] "lACA(R)"           "EB"                "PB(R1)"           
#> [151] "FBl1"             
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "EBr3d"             "AL-VC4(R)"         "AL-DL5(R)"        
#>   [4] "AL-VC2(R)"         "b'2(R)"            "bL(L)"            
#>   [7] "AL-DL4(R)"         "RUB(R)"            "GC"               
#>  [10] "aL(R)"             "AB(R)"             "INP"              
#>  [13] "AL-DC3(R)"         "EB"                "PB(L6)"           
#>  [16] "ICL(L)"            "FBl6"              "AL-DL1(R)"        
#>  [19] "gL(L)"             "AL-DM5(L)"         "AL-VA6(L)"        
#>  [22] "EPA(L)"            "PB(L8)"            "ATL(R)"           
#>  [25] "AME(R)"            "PB(R7)"            "AL-DL2v(R)"       
#>  [28] "b'L(R)"            "AL-DA1(R)"         "AL-DP1l(R)"       
#>  [31] "PB(R5)"            "PB(R1)"            "a'1(R)"           
#>  [34] "SAD"               "CRE(R)"            "AL-DC2(L)"        
#>  [37] "FBl9"              "b1(R)"             "AL-VC5(R)"        
#>  [40] "PB(R3)"            "AL-DA4l(R)"        "AOTU(R)"          
#>  [43] "AL-VL1(R)"         "LX(R)"             "PB(L7)"           
#>  [46] "OL(R)"             "EBr6"              "EBr2r4"           
#>  [49] "AL-VM3(R)"         "lACA(R)"           "g2(R)"            
#>  [52] "AL-VA1d(R)"        "AL-DM6(R)"         "b'1(R)"           
#>  [55] "AL-D(L)"           "SIP(R)"            "RUB(L)"           
#>  [58] "PB(L4)"            "AL-DM1(R)"         "a3(R)"            
#>  [61] "PRW"               "AL-DL4(L)"         "PB(R9)"           
#>  [64] "AL(L)"             "CA(L)"             "g4(R)"            
#>  [67] "MB(+ACA)(R)"       "bL(R)"             "AL-DL2d(R)"       
#>  [70] "SCL(L)"            "mALT(R)"           "SCL(R)"           
#>  [73] "SNP(R)"            "AMMC"              "a1(R)"            
#>  [76] "PED(R)"            "NO3(R)"            "AL-DM2(R)"        
#>  [79] "AL-DP1m(R)"        "CRE(-ROB,-RUB)(R)" "PVLP(R)"          
#>  [82] "a'L(L)"            "AL(R)"             "PLP(R)"           
#>  [85] "GOR(R)"            "AL-DA4m(R)"        "AL-V(R)"          
#>  [88] "AOT(R)"            "AL-VM7v(L)"        "gL(R)"            
#>  [91] "FB-column3"        "EBr3am"            "PENP"             
#>  [94] "b2(R)"             "ICL(R)"            "NO(L)"            
#>  [97] "POC"               "g1(R)"             "ME(R)"            
#> [100] "NO"                "AL-VL2a(R)"        "PB(L5)"           
#> [103] "WED(R)"            "AL-VM7d(L)"        "FBl7"             
#> [106] "AL-VM5v(R)"        "LOP(R)"            "dACA(R)"          
#> [109] "FBl1"              "b'L(L)"            "PB(L3)"           
#> [112] "VMNP"              "vACA(R)"           "AL-DP1m(L)"       
#> [115] "CRE(-RUB)(L)"      "CAN(R)"            "NO3(L)"           
#> [118] "AL-DA2(R)"         "AL-VM7v(R)"        "VLNP(R)"          
#> [121] "SPS(L)"            "AL-DC4(L)"         "AL-DM4(L)"        
#> [124] "GOR(L)"            "ATL(L)"            "AL-DM6(L)"        
#> [127] "LH(R)"             "MB(R)"             "AL-DC1(L)"        
#> [130] "mALT(L)"           "NO(R)"             "FBl8"             
#> [133] "hemibrain"         "PB(R8)"            "BU(R)"            
#> [136] "AL-VA5(R)"         "AL-VP4(R)"         "PB(R6)"           
#> [139] "LAL(-GA)(R)"       "PB(R2)"            "AL-VL2p(R)"       
#> [142] "PB(L1)"            "g5(R)"             "AL-VA1v(R)"       
#> [145] "AL-DM1(L)"         "PB(L9)"            "FBl3"             
#> [148] "aL(L)"             "AL-VA3(R)"         "SPS(R)"           
#> [151] "FLA(R)"            "NO1(R)"            "AL-VP1l(R)"       
#> [154] "AL-DA3(R)"         "AL-VA2(R)"         "AL-VC3m(R)"       
#> [157] "MB(L)"             "SAD(-AMMC)"        "CX"               
#> [160] "a2(R)"             "LAL(L)"            "AL-DM3(R)"        
#> [163] "AL-VM5d(R)"        "PB"                "AL-DA2(L)"        
#> [166] "VES(L)"            "EBr1"              "EBr5"             
#> [169] "AL-VP1m(R)"        "AL-DM4(R)"         "a'3(R)"           
#> [172] "AL-DL5(L)"         "FBl5"              "LAL(R)"           
#> [175] "AL-DC1(R)"         "AVLP(R)"           "AL-DC2(R)"        
#> [178] "AL-VP3(R)"         "IPS(R)"            "NO2(R)"           
#> [181] "AL-VA6(R)"         "NO1(L)"            "a'L(R)"           
#> [184] "AL-DA3(L)"         "AL-VP2(R)"         "PB(R4)"           
#> [187] "AL-DC4(R)"         "AL-VM2(R)"         "ROB(R)"           
#> [190] "AL-VA7l(R)"        "SMP(R)"            "AL-DM3(L)"        
#> [193] "CA(R)"             "LX(L)"             "AL-VM7d(R)"       
#> [196] "a'2(R)"            "EPA(R)"            "AL-DA4m(L)"       
#> [199] "g3(R)"             "LO(R)"             "SLP(R)"           
#> [202] "SIP(L)"            "AB(L)"             "IB"               
#> [205] "CRE(L)"            "SNP(L)"            "AL-VM1(R)"        
#> [208] "GA(R)"             "AL-VA7m(R)"        "EBr3pw"           
#> [211] "GNG"               "AL-VC3l(R)"        "AL-VP1d(R)"       
#> [214] "SMP(L)"            "AL-DL3(R)"         "AL-VA4(R)"        
#> [217] "FBl4"              "AL-VM4(R)"         "GF(R)"            
#> [220] "BU(L)"             "VES(R)"            "AL-DM5(R)"        
#> [223] "AL-VP5(R)"         "NO2(L)"            "AL-DM2(L)"        
#> [226] "AL-VC1(R)"         "PB(L2)"            "FB"               
#> [229] "FBl2"              "AL-D(R)"          
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "b'2(R)"            "EBr6"              "bL(R)"            
#>   [4] "AL-VA7l(R)"        "AL-VP1l(R)"        "b'1(R)"           
#>   [7] "b2(R)"             "PLP(R)"            "a'L(R)"           
#>  [10] "AL(L)"             "FBl9"              "AL-DC2(R)"        
#>  [13] "PED(R)"            "EBr5"              "a1(R)"            
#>  [16] "CRE(R)"            "AL-DC2(L)"         "bL(L)"            
#>  [19] "AL-DA1(R)"         "AL-VM5d(R)"        "b'L(R)"           
#>  [22] "AL-DM5(L)"         "EBr2r4"            "ROB(R)"           
#>  [25] "FBl7"              "OL(R)"             "AL-VM1(R)"        
#>  [28] "a'3(R)"            "SLP(R)"            "AL-VA1v(R)"       
#>  [31] "SCL(L)"            "CRE(-ROB,-RUB)(R)" "AOTU(R)"          
#>  [34] "SAD(-AMMC)"        "FBl3"              "AL-DA4m(R)"       
#>  [37] "AL-VL2a(R)"        "IB"                "CX"               
#>  [40] "a'2(R)"            "AL-VM7d(R)"        "LOP(R)"           
#>  [43] "GOR(R)"            "AL-VC3m(R)"        "AL-VL2p(R)"       
#>  [46] "PB(L4)"            "AL-VP1m(R)"        "AL-VM5v(R)"       
#>  [49] "AL-DC3(R)"         "AL-D(R)"           "AL-VA4(R)"        
#>  [52] "FLA(R)"            "PB(L9)"            "FBl6"             
#>  [55] "ATL(R)"            "AL-DM6(R)"         "PRW"              
#>  [58] "PB(R9)"            "AL-DA4l(R)"        "AL-VM2(R)"        
#>  [61] "AL-DM3(L)"         "AL-DP1m(L)"        "lACA(R)"          
#>  [64] "PB(R1)"            "WED(R)"            "EBr3pw"           
#>  [67] "AL-DA3(R)"         "AL-VA1d(R)"        "AL-VC3l(R)"       
#>  [70] "mALT(L)"           "gL(L)"             "gL(R)"            
#>  [73] "FBl5"              "AL-VP1d(R)"        "ATL(L)"           
#>  [76] "LH(R)"             "CAN(R)"            "PB(R4)"           
#>  [79] "AL-DC4(R)"         "AL-VL1(R)"         "FB"               
#>  [82] "AL-D(L)"           "FBl4"              "aL(L)"            
#>  [85] "AL-DM4(R)"         "EBr3am"            "PB(R6)"           
#>  [88] "a2(R)"             "FBl2"              "IPS(R)"           
#>  [91] "AL-VC5(R)"         "AL-VM4(R)"         "NO1(L)"           
#>  [94] "SPS(L)"            "PVLP(R)"           "AL-VA5(R)"        
#>  [97] "g1(R)"             "ME(R)"             "AL-DM3(R)"        
#> [100] "g3(R)"             "GA(R)"             "MB(L)"            
#> [103] "FBl8"              "ICL(R)"            "AME(R)"           
#> [106] "RUB(L)"            "RUB(R)"            "LAL(-GA)(R)"      
#> [109] "AL-DM4(L)"         "FB-column3"        "b1(R)"            
#> [112] "NO2(R)"            "a3(R)"             "AL-DP1l(R)"       
#> [115] "g4(R)"             "MB(+ACA)(R)"       "AL-DL2v(R)"       
#> [118] "LAL(L)"            "AL-VP4(R)"         "PB(R8)"           
#> [121] "SIP(R)"            "PB(L2)"            "CA(L)"            
#> [124] "GC"                "SCL(R)"            "BU(L)"            
#> [127] "PB(L1)"            "AL-DA2(L)"         "AL-VM7v(R)"       
#> [130] "PB(R5)"            "AMMC"              "BU(R)"            
#> [133] "NO(L)"             "LAL(R)"            "MB(R)"            
#> [136] "AL-V(R)"           "NO3(R)"            "SMP(R)"           
#> [139] "CRE(-RUB)(L)"      "FBl1"              "AL-DA2(R)"        
#> [142] "VMNP"              "AVLP(R)"           "AL-VA3(R)"        
#> [145] "AL-VP5(R)"         "AL-VP3(R)"         "PB(L3)"           
#> [148] "SAD"               "PB(L8)"            "GF(R)"            
#> [151] "a'L(L)"            "vACA(R)"           "AL-DM1(L)"        
#> [154] "SNP(L)"            "PB(R2)"            "INP"              
#> [157] "AL-DC1(R)"         "AL-DL5(R)"         "AL-VM3(R)"        
#> [160] "NO1(R)"            "PB"                "EBr3d"            
#> [163] "PB(L6)"            "LX(L)"             "AL-DM5(R)"        
#> [166] "NO3(L)"            "PB(R3)"            "NO2(L)"           
#> [169] "EB"                "VES(R)"            "AL-VA2(R)"        
#> [172] "g2(R)"             "LX(R)"             "b'L(L)"           
#> [175] "AL-VP2(R)"         "dACA(R)"           "LO(R)"            
#> [178] "AL-VC2(R)"         "SPS(R)"            "GOR(L)"           
#> [181] "SMP(L)"            "AL-DC1(L)"         "AL-DC4(L)"        
#> [184] "EPA(L)"            "AL-DL4(R)"         "AL-DM2(R)"        
#> [187] "AL-VC1(R)"         "EPA(R)"            "AL-DA4m(L)"       
#> [190] "NO(R)"             "AL(R)"             "PB(L5)"           
#> [193] "g5(R)"             "CA(R)"             "aL(R)"            
#> [196] "CRE(L)"            "GNG"               "PB(R7)"           
#> [199] "AL-DA3(L)"         "a'1(R)"            "AL-VC4(R)"        
#> [202] "AOT(R)"            "mALT(R)"           "AL-DL5(L)"        
#> [205] "NO"                "AL-DL2d(R)"        "AL-VA7m(R)"       
#> [208] "AL-VM7d(L)"        "PB(L7)"            "AL-DM6(L)"        
#> [211] "AL-VM7v(L)"        "ICL(L)"            "AL-DL1(R)"        
#> [214] "AL-DP1m(R)"        "AL-DM2(L)"         "VES(L)"           
#> [217] "AL-DL3(R)"         "AL-VA6(R)"         "SIP(L)"           
#> [220] "SNP(R)"            "AL-DM1(R)"         "AB(L)"            
#> [223] "PENP"              "EBr1"              "AL-DL4(L)"        
#> [226] "AL-VA6(L)"         "AB(R)"             "VLNP(R)"          
#> [229] "POC"               "hemibrain"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`manc:v1.0`
#> $`manc:v1.0`$`last-mod`
#> [1] "2023-05-31 23:54:01"
#> 
#> $`manc:v1.0`$uuid
#> [1] "59b37970bc7a4341b9a3a965a0d6b402"
#> 
#> $`manc:v1.0`$ROIs
#>  [1] "MesoAN(L)"        "LegNp(T1)(R)"     "HTct(UTct-T3)(R)" "GF(L)"           
#>  [5] "VProN(R)"         "PrN(L)"           "LegNp(T2)(L)"     "mVAC(T2)(L)"     
#>  [9] "CvN(L)"           "WTct(UTct-T2)(L)" "AbN1(L)"          "mVAC(T2)(R)"     
#> [13] "AbN1(R)"          "ProCN(L)"         "MetaLN(R)"        "mVAC(T3)(R)"     
#> [17] "IntTct"           "AbN4(R)"          "MesoLN(R)"        "DMetaN(R)"       
#> [21] "MetaLN(L)"        "PDMN(L)"          "mVAC(T1)(L)"      "AbN2(L)"         
#> [25] "DMetaN(L)"        "AbN2(R)"          "DProN(L)"         "ProAN(R)"        
#> [29] "ProLN(R)"         "LegNp(T2)(R)"     "Ov(L)"            "WTct(UTct-T2)(R)"
#> [33] "ProAN(L)"         "NTct(UTct-T1)(L)" "NTct(UTct-T1)(R)" "CvN(R)"          
#> [37] "LegNp(T3)(R)"     "DProN(R)"         "VProN(L)"         "PrN(R)"          
#> [41] "mVAC(T1)(R)"      "AbN3(R)"          "PDMN(R)"          "mVAC(T3)(L)"     
#> [45] "Ov(R)"            "LTct"             "GF(R)"            "AbN4(L)"         
#> [49] "AbNT"             "MesoLN(L)"        "CV"               "ANm"             
#> [53] "ADMN(R)"          "ProCN(R)"         "LegNp(T1)(L)"     "LegNp(T3)(L)"    
#> [57] "HTct(UTct-T3)(L)" "ProLN(L)"         "ADMN(L)"          "AbN3(L)"         
#> [61] "MesoAN(R)"       
#> 
#> $`manc:v1.0`$superLevelROIs
#>  [1] "ADMN(L)"          "ADMN(R)"          "Ov(L)"            "Ov(R)"           
#>  [5] "ANm"              "AbN1(L)"          "AbN1(R)"          "AbN2(L)"         
#>  [9] "AbN2(R)"          "AbN3(L)"          "AbN3(R)"          "AbN4(L)"         
#> [13] "AbN4(R)"          "AbNT"             "CV"               "CvN(L)"          
#> [17] "CvN(R)"           "DMetaN(L)"        "DMetaN(R)"        "DProN(L)"        
#> [21] "DProN(R)"         "GF(L)"            "GF(R)"            "HTct(UTct-T3)(L)"
#> [25] "HTct(UTct-T3)(R)" "LegNp(T1)(L)"     "LegNp(T1)(R)"     "LegNp(T2)(L)"    
#> [29] "LegNp(T2)(R)"     "LegNp(T3)(L)"     "LegNp(T3)(R)"     "IntTct"          
#> [33] "LTct"             "MesoAN(L)"        "MesoAN(R)"        "MesoLN(L)"       
#> [37] "MesoLN(R)"        "MetaLN(L)"        "MetaLN(R)"        "NTct(UTct-T1)(L)"
#> [41] "NTct(UTct-T1)(R)" "PDMN(L)"          "PrN(L)"           "PrN(R)"          
#> [45] "ProCN(L)"         "ProCN(R)"         "PDMN(R)"          "ProAN(L)"        
#> [49] "ProAN(R)"         "ProLN(L)"         "ProLN(R)"         "VProN(L)"        
#> [53] "VProN(R)"         "WTct(UTct-T2)(L)" "WTct(UTct-T2)(R)" "mVAC(T1)(L)"     
#> [57] "mVAC(T1)(R)"      "mVAC(T2)(L)"      "mVAC(T2)(R)"      "mVAC(T3)(L)"     
#> [61] "mVAC(T3)(R)"     
#> 
#> $`manc:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/manc-connectome"
#> 
#> $`manc:v1.0`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                           
#> [1,] "POST" "/api/v:ver/npexplorer/neuronmeta"             
#> [2,] "GET"  "/api/npexplorer/autapses"                     
#> [3,] "POST" "/api/npexplorer/distribution"                 
#> [4,] "POST" "/api/v:ver/roimeshes/mesh/:dataset/:roi"      
#> [5,] "GET"  "/*"                                           
#> [6,] "GET"  "/api/v:ver/npexplorer/celltype/:dataset/:type"
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"    
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [5,] "github.com/labstack/echo/v4.StaticDirectoryHandler.func1"                                  
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }