Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "PB(R1)"            "EBr6"              "EBr2r4"           
#>   [4] "AL-DC1(R)"         "AL-VL2a(R)"        "ROB(R)"           
#>   [7] "FBl4"              "PB(L2)"            "PB(R3)"           
#>  [10] "VES(R)"            "AL-VA5(R)"         "lACA(R)"          
#>  [13] "AL-DA2(L)"         "AL-DP1m(L)"        "PB(L6)"           
#>  [16] "NO1(L)"            "AL-DM3(L)"         "AL-VA7l(R)"       
#>  [19] "CRE(L)"            "AL-DL4(R)"         "AL-VM5v(R)"       
#>  [22] "SMP(R)"            "g5(R)"             "SAD(-AMMC)"       
#>  [25] "AL-VP5(R)"         "AL-VP1d(R)"        "RUB(R)"           
#>  [28] "AL-DC2(R)"         "AL-VC2(R)"         "g1(R)"            
#>  [31] "EBr5"              "SPS(L)"            "VMNP"             
#>  [34] "AL-DM5(R)"         "AL-VM7d(R)"        "NO2(L)"           
#>  [37] "FB"                "AL-DM5(L)"         "FBl1"             
#>  [40] "LX(L)"             "hemibrain"         "EB"               
#>  [43] "AL-DL5(R)"         "a'1(R)"            "AOTU(R)"          
#>  [46] "EBr3d"             "IB"                "SMP(L)"           
#>  [49] "MB(R)"             "AL-DA4l(R)"        "mALT(L)"          
#>  [52] "PRW"               "a'2(R)"            "SIP(R)"           
#>  [55] "AL-VM5d(R)"        "BU(R)"             "PVLP(R)"          
#>  [58] "AL-VM4(R)"         "AL-VP1m(R)"        "AL-VC4(R)"        
#>  [61] "vACA(R)"           "NO1(R)"            "PLP(R)"           
#>  [64] "gL(L)"             "AL-DA3(R)"         "g3(R)"            
#>  [67] "IB112"             "LAL(R)"            "ATL(L)"           
#>  [70] "SPS(R)"            "AL-DC3(R)"         "AL-VA1v(R)"       
#>  [73] "AL-VL2p(R)"        "AL(L)"             "AL-DC4(R)"        
#>  [76] "OL(R)"             "AL-VA6(R)"         "AL-VC3m(R)"       
#>  [79] "AL-VM1(R)"         "aL(R)"             "EBr1"             
#>  [82] "PB(L1)"            "PB(L4)"            "AL-VP3(R)"        
#>  [85] "AL-VP2(R)"         "AB(L)"             "CA(L)"            
#>  [88] "AL-DM1(L)"         "PB(R2)"            "IPS(R)"           
#>  [91] "AL-DP1l(R)"        "LX(R)"             "AL-VC3l(R)"       
#>  [94] "NO3(R)"            "AL-D(L)"           "AL-VM7v(L)"       
#>  [97] "NO3(L)"            "CX"                "GA(R)"            
#> [100] "g4(R)"             "b'L(L)"            "AL-VA4(R)"        
#> [103] "b'1(R)"            "GF(R)"             "AL-VM7d(L)"       
#> [106] "FBl7"              "EPA(L)"            "FLA(R)"           
#> [109] "AL-DC4(L)"         "AL-DM2(L)"         "WED(R)"           
#> [112] "EBr3am"            "b1(R)"             "AL-VA3(R)"        
#> [115] "POC"               "NO2(R)"            "PB(R4)"           
#> [118] "AL-DL2v(R)"        "MB(L)"             "PB(L8)"           
#> [121] "AL-DM4(L)"         "ICL(L)"            "NO(R)"            
#> [124] "bL(R)"             "AL-V(R)"           "a'3(R)"           
#> [127] "AME(R)"            "AL-DL2d(R)"        "AL-DL3(R)"        
#> [130] "bL(L)"             "CA(R)"             "SCL(L)"           
#> [133] "AL-VA6(L)"         "PB(L9)"            "CRE(-RUB)(L)"     
#> [136] "GNG"               "PB(R7)"            "PB(L7)"           
#> [139] "a2(R)"             "AB(R)"             "AL-VM7v(R)"       
#> [142] "LH(R)"             "AL-VL1(R)"         "ICL(R)"           
#> [145] "a3(R)"             "SAD"               "SIP(L)"           
#> [148] "LAL(-GA)(R)"       "MB(+ACA)(R)"       "LAL(L)"           
#> [151] "INP"               "AL(R)"             "VLNP(R)"          
#> [154] "PB"                "ME(R)"             "PB(R9)"           
#> [157] "AL-D(R)"           "GC"                "SNP(L)"           
#> [160] "RUB(L)"            "AL-DP1m(R)"        "PENP"             
#> [163] "BU(L)"             "AL-DM6(L)"         "NO(L)"            
#> [166] "AL-DM2(R)"         "AL-VA1d(R)"        "mALT(R)"          
#> [169] "AL-DA3(L)"         "FB-column3"        "AL-DL1(R)"        
#> [172] "CAN(R)"            "AL-DM4(R)"         "PB(L3)"           
#> [175] "AL-DM3(R)"         "g2(R)"             "FBl3"             
#> [178] "SLP(R)"            "AL-DA1(R)"         "SNP(R)"           
#> [181] "AL-DM6(R)"         "AL-VM3(R)"         "PB(L5)"           
#> [184] "PED(R)"            "AL-VC1(R)"         "CRE(-ROB,-RUB)(R)"
#> [187] "FBl6"              "VES(L)"            "AL-VP1l(R)"       
#> [190] "a1(R)"             "b2(R)"             "AL-DA4m(L)"       
#> [193] "AL-DL5(L)"         "FBl5"              "FBl2"             
#> [196] "FBl8"              "AL-VC5(R)"         "b'2(R)"           
#> [199] "PB(R8)"            "a'L(R)"            "gL(R)"            
#> [202] "AMMC"              "CRE(R)"            "SCL(R)"           
#> [205] "AL-DA4m(R)"        "AL-VA7m(R)"        "GOR(L)"           
#> [208] "b'L(R)"            "AL-DC1(L)"         "AL-VP4(R)"        
#> [211] "AL-DC2(L)"         "AVLP(R)"           "aL(L)"            
#> [214] "EPA(R)"            "PB(R5)"            "GOR(R)"           
#> [217] "AL-VM2(R)"         "LOP(R)"            "PB(R6)"           
#> [220] "dACA(R)"           "NO"                "ATL(R)"           
#> [223] "EBr3pw"            "a'L(L)"            "AL-DA2(R)"        
#> [226] "AL-VA2(R)"         "AOT(R)"            "AL-DL4(L)"        
#> [229] "LO(R)"             "FBl9"              "AL-DM1(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "EBr6"              "PB(R2)"            "AME(R)"           
#>   [4] "MB(R)"             "NO(L)"             "vACA(R)"          
#>   [7] "NO1(L)"            "MB(L)"             "ME(R)"            
#>  [10] "PB(L6)"            "PLP(R)"            "b1(R)"            
#>  [13] "PB(L8)"            "EB"                "g5(R)"            
#>  [16] "a'L(L)"            "PB(L3)"            "b'L(R)"           
#>  [19] "FBl7"              "AB(R)"             "EBr5"             
#>  [22] "a1(R)"             "PB(R4)"            "SIP(L)"           
#>  [25] "a'L(R)"            "FBl1"              "RUB(L)"           
#>  [28] "VES(L)"            "b'1(R)"            "a2(R)"            
#>  [31] "SPS(R)"            "WED(R)"            "FBl5"             
#>  [34] "a3(R)"             "LAL(-GA)(R)"       "dACA(R)"          
#>  [37] "PB"                "CRE(L)"            "OL(R)"            
#>  [40] "PB(L5)"            "AMMC"              "EPA(R)"           
#>  [43] "PB(R1)"            "NO2(R)"            "GOR(R)"           
#>  [46] "ICL(L)"            "CA(R)"             "PB(L2)"           
#>  [49] "FB"                "ATL(R)"            "ICL(R)"           
#>  [52] "INP"               "PB(L4)"            "PENP"             
#>  [55] "GOR(L)"            "EBr2r4"            "PB(L7)"           
#>  [58] "a'3(R)"            "PB(R6)"            "AOTU(R)"          
#>  [61] "SAD"               "bL(R)"             "NO(R)"            
#>  [64] "AB(L)"             "CRE(-ROB,-RUB)(R)" "mALT(R)"          
#>  [67] "CRE(R)"            "a'2(R)"            "LOP(R)"           
#>  [70] "mALT(L)"           "PED(R)"            "BU(R)"            
#>  [73] "LO(R)"             "LH(R)"             "EBr3pw"           
#>  [76] "AOT(R)"            "PB(R5)"            "BU(L)"            
#>  [79] "NO"                "PB(L9)"            "NO3(L)"           
#>  [82] "AL(L)"             "SCL(L)"            "PRW"              
#>  [85] "b2(R)"             "PB(L1)"            "PB(R8)"           
#>  [88] "LX(R)"             "PVLP(R)"           "FBl8"             
#>  [91] "aL(L)"             "lACA(R)"           "a'1(R)"           
#>  [94] "NO2(L)"            "FBl9"              "CAN(R)"           
#>  [97] "SMP(R)"            "EBr3am"            "EBr3d"            
#> [100] "PB(R3)"            "EPA(L)"            "gL(L)"            
#> [103] "ROB(R)"            "FBl3"              "SMP(L)"           
#> [106] "g1(R)"             "SCL(R)"            "EBr1"             
#> [109] "AVLP(R)"           "LX(L)"             "AL(R)"            
#> [112] "VLNP(R)"           "SNP(R)"            "SNP(L)"           
#> [115] "FBl4"              "FBl6"              "IB"               
#> [118] "ATL(L)"            "CX"                "RUB(R)"           
#> [121] "hemibrain"         "PB(R7)"            "SPS(L)"           
#> [124] "IPS(R)"            "gL(R)"             "GNG"              
#> [127] "b'L(L)"            "CA(L)"             "GA(R)"            
#> [130] "FLA(R)"            "VMNP"              "SIP(R)"           
#> [133] "VES(R)"            "LAL(R)"            "b'2(R)"           
#> [136] "LAL(L)"            "NO1(R)"            "PB(R9)"           
#> [139] "bL(L)"             "GF(R)"             "g3(R)"            
#> [142] "GC"                "g4(R)"             "MB(+ACA)(R)"      
#> [145] "FBl2"              "POC"               "g2(R)"            
#> [148] "SLP(R)"            "aL(R)"             "NO3(R)"           
#> [151] "SAD(-AMMC)"       
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "PB(R8)"            "PB"                "a'1(R)"           
#>   [4] "CRE(-ROB,-RUB)(R)" "NO1(R)"            "ATL(R)"           
#>   [7] "AL-DC2(R)"         "AL-VC3m(R)"        "POC"              
#>  [10] "EBr6"              "EBr3d"             "GOR(R)"           
#>  [13] "a2(R)"             "PB(R7)"            "AL-DM6(L)"        
#>  [16] "AL-VA1d(R)"        "AL-DM1(L)"         "NO"               
#>  [19] "AL-VC3l(R)"        "AL-VC4(R)"         "SMP(R)"           
#>  [22] "AL-VP2(R)"         "IB"                "PB(L3)"           
#>  [25] "AL-DA4m(L)"        "FBl9"              "LAL(R)"           
#>  [28] "AL-VA4(R)"         "g1(R)"             "NO1(L)"           
#>  [31] "AL-DC3(R)"         "NO2(L)"            "BU(R)"            
#>  [34] "AL-DM4(L)"         "AL-DA2(R)"         "AL-DL3(R)"        
#>  [37] "AL-DM3(R)"         "PB(L7)"            "ATL(L)"           
#>  [40] "AL-VM7v(R)"        "AL-DL4(L)"         "AL-DM3(L)"        
#>  [43] "CA(R)"             "RUB(L)"            "AL-VM1(R)"        
#>  [46] "AL-VM2(R)"         "PB(R5)"            "LAL(L)"           
#>  [49] "ROB(R)"            "CRE(L)"            "AL-VP1m(R)"       
#>  [52] "AL-DA2(L)"         "a'L(L)"            "AL-VM3(R)"        
#>  [55] "AL-VP1d(R)"        "AB(L)"             "PENP"             
#>  [58] "FLA(R)"            "SNP(L)"            "LO(R)"            
#>  [61] "a3(R)"             "mALT(L)"           "PB(R1)"           
#>  [64] "AL-DM6(R)"         "AL-D(R)"           "AL-VL2p(R)"       
#>  [67] "RUB(R)"            "CX"                "EPA(R)"           
#>  [70] "FBl4"              "AB(R)"             "AL-DC4(R)"        
#>  [73] "AL-VM7d(R)"        "PED(R)"            "aL(L)"            
#>  [76] "PB(R4)"            "SCL(L)"            "PB(R6)"           
#>  [79] "AL-VA6(L)"         "AL-DL2d(R)"        "AL-VM5v(R)"       
#>  [82] "NO3(R)"            "PB(L1)"            "AVLP(R)"          
#>  [85] "AL-DP1m(R)"        "FB-column3"        "FBl8"             
#>  [88] "b'L(L)"            "AL-DA3(R)"         "SLP(R)"           
#>  [91] "AL-DC2(L)"         "AL-VA7m(R)"        "AL-VM4(R)"        
#>  [94] "CA(L)"             "GOR(L)"            "EBr5"             
#>  [97] "g4(R)"             "AL-DM4(R)"         "VLNP(R)"          
#> [100] "PB(R2)"            "AL-V(R)"           "ME(R)"            
#> [103] "AL-VC2(R)"         "EBr1"              "PB(L9)"           
#> [106] "LX(L)"             "EPA(L)"            "AL-VP3(R)"        
#> [109] "IPS(R)"            "AL-VC5(R)"         "AL-VL1(R)"        
#> [112] "SPS(L)"            "AL-DC4(L)"         "PB(L4)"           
#> [115] "VES(L)"            "AL-DP1l(R)"        "AL(R)"            
#> [118] "AL-VA2(R)"         "GF(R)"             "PB(L8)"           
#> [121] "mALT(R)"           "AL-DA3(L)"         "g3(R)"            
#> [124] "AOTU(R)"           "lACA(R)"           "CRE(R)"           
#> [127] "PLP(R)"            "AL-VM7d(L)"        "GNG"              
#> [130] "b'2(R)"            "b2(R)"             "AL-DC1(L)"        
#> [133] "vACA(R)"           "aL(R)"             "EBr2r4"           
#> [136] "ICL(R)"            "AL-VA1v(R)"        "AL-VA6(R)"        
#> [139] "SIP(R)"            "AL-DC1(R)"         "b'1(R)"           
#> [142] "AL-VA3(R)"         "AL-VL2a(R)"        "AL-VP1l(R)"       
#> [145] "SNP(R)"            "gL(R)"             "FBl6"             
#> [148] "NO(L)"             "AL-DM2(R)"         "a1(R)"            
#> [151] "g2(R)"             "PB(R9)"            "g5(R)"            
#> [154] "EB"                "bL(L)"             "NO3(L)"           
#> [157] "EBr3pw"            "a'3(R)"            "hemibrain"        
#> [160] "PB(R3)"            "AL-VM5d(R)"        "dACA(R)"          
#> [163] "FBl5"              "AL-DL2v(R)"        "FBl2"             
#> [166] "AL-VC1(R)"         "AOT(R)"            "WED(R)"           
#> [169] "PVLP(R)"           "AL-DA4m(R)"        "PRW"              
#> [172] "VMNP"              "EBr3am"            "SAD"              
#> [175] "NO(R)"             "AL-DL5(R)"         "AL-DL1(R)"        
#> [178] "SMP(L)"            "SCL(R)"            "AL-DM2(L)"        
#> [181] "CRE(-RUB)(L)"      "OL(R)"             "BU(L)"            
#> [184] "a'L(R)"            "gL(L)"             "bL(R)"            
#> [187] "AL-DM5(R)"         "GC"                "MB(+ACA)(R)"      
#> [190] "AL-DA1(R)"         "LX(R)"             "AL-DL5(L)"        
#> [193] "AL-DM1(R)"         "LAL(-GA)(R)"       "CAN(R)"           
#> [196] "AMMC"              "MB(L)"             "AL-DP1m(L)"       
#> [199] "FBl1"              "FBl3"              "ICL(L)"           
#> [202] "VES(R)"            "INP"               "PB(L2)"           
#> [205] "GA(R)"             "SIP(L)"            "FBl7"             
#> [208] "AL-VP5(R)"         "NO2(R)"            "FB"               
#> [211] "b'L(R)"            "AL-D(L)"           "AL-DM5(L)"        
#> [214] "AL-VP4(R)"         "AL-VA7l(R)"        "SPS(R)"           
#> [217] "AL(L)"             "PB(L6)"            "SAD(-AMMC)"       
#> [220] "b1(R)"             "AL-VA5(R)"         "LOP(R)"           
#> [223] "PB(L5)"            "a'2(R)"            "LH(R)"            
#> [226] "AME(R)"            "AL-DA4l(R)"        "AL-VM7v(L)"       
#> [229] "MB(R)"             "AL-DL4(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "ICL(R)"            "AL-DL4(L)"         "SIP(R)"           
#>   [4] "NO"                "g5(R)"             "AVLP(R)"          
#>   [7] "LX(L)"             "AL-DL2d(R)"        "EBr2r4"           
#>  [10] "a'L(R)"            "gL(L)"             "AL-DL5(L)"        
#>  [13] "AL-DM4(L)"         "FBl1"              "AL(R)"            
#>  [16] "AL-VM7v(R)"        "PRW"               "CRE(R)"           
#>  [19] "PB(L5)"            "LH(R)"             "AL-DM2(R)"        
#>  [22] "AL-VM2(R)"         "SCL(L)"            "a1(R)"            
#>  [25] "FB"                "CRE(L)"            "VMNP"             
#>  [28] "MB(+ACA)(R)"       "PB(R3)"            "RUB(L)"           
#>  [31] "SNP(R)"            "AL-DA3(L)"         "g4(R)"            
#>  [34] "AL(L)"             "AL-DM4(R)"         "AL-VA4(R)"        
#>  [37] "PB(R8)"            "EBr3d"             "a'1(R)"           
#>  [40] "mALT(L)"           "a2(R)"             "LX(R)"            
#>  [43] "FLA(R)"            "AL-VA6(L)"         "ME(R)"            
#>  [46] "ATL(L)"            "PVLP(R)"           "AL-DA1(R)"        
#>  [49] "AL-VP3(R)"         "CRE(-ROB,-RUB)(R)" "SIP(L)"           
#>  [52] "AL-DM5(L)"         "AL-VA1v(R)"        "b'2(R)"           
#>  [55] "PB(L7)"            "AL-VC5(R)"         "b'1(R)"           
#>  [58] "AL-VA1d(R)"        "PB(L8)"            "FBl6"             
#>  [61] "GOR(R)"            "AL-VA3(R)"         "AL-DM1(L)"        
#>  [64] "AL-VM4(R)"         "VLNP(R)"           "AMMC"             
#>  [67] "CA(L)"             "b'L(R)"            "EB"               
#>  [70] "SAD(-AMMC)"        "AL-VA6(R)"         "a'2(R)"           
#>  [73] "CRE(-RUB)(L)"      "AB(R)"             "AL-VP5(R)"        
#>  [76] "LAL(L)"            "dACA(R)"           "AL-DL3(R)"        
#>  [79] "GF(R)"             "AL-DA3(R)"         "AL-DA4l(R)"       
#>  [82] "AL-DL1(R)"         "AL-VA7m(R)"        "AL-VP1d(R)"       
#>  [85] "AL-VC4(R)"         "EBr1"              "AME(R)"           
#>  [88] "bL(L)"             "AL-VM3(R)"         "AL-DA2(R)"        
#>  [91] "GNG"               "PB(L4)"            "b1(R)"            
#>  [94] "LAL(R)"            "AL-VC1(R)"         "AL-VP4(R)"        
#>  [97] "EBr6"              "CA(R)"             "AL-DM2(L)"        
#> [100] "FBl9"              "PB(R4)"            "PB(R6)"           
#> [103] "aL(R)"             "AL-VC3l(R)"        "AL-VL1(R)"        
#> [106] "POC"               "g3(R)"             "hemibrain"        
#> [109] "FBl4"              "AL-DC1(R)"         "AL-VL2a(R)"       
#> [112] "lACA(R)"           "NO1(R)"            "FB-column3"       
#> [115] "OL(R)"             "BU(R)"             "AL-VL2p(R)"       
#> [118] "aL(L)"             "VES(L)"            "AL-DC2(R)"        
#> [121] "MB(L)"             "AL-DC2(L)"         "EBr3am"           
#> [124] "PENP"              "b2(R)"             "MB(R)"            
#> [127] "AL-VM1(R)"         "a'3(R)"            "vACA(R)"          
#> [130] "AOTU(R)"           "PB(R2)"            "AOT(R)"           
#> [133] "GA(R)"             "b'L(L)"            "AL-DA4m(R)"       
#> [136] "AL-VM5v(R)"        "AL-VP1l(R)"        "PB(L3)"           
#> [139] "AL-DM5(R)"         "FBl2"              "EPA(L)"           
#> [142] "NO(L)"             "AL-DC4(R)"         "LOP(R)"           
#> [145] "NO1(L)"            "NO3(L)"            "AL-DM6(L)"        
#> [148] "NO(R)"             "NO2(R)"            "AL-DL5(R)"        
#> [151] "SAD"               "PLP(R)"            "AL-DL4(R)"        
#> [154] "AL-DP1m(R)"        "GC"                "AL-DP1m(L)"       
#> [157] "AL-VA5(R)"         "IB"                "BU(L)"            
#> [160] "LO(R)"             "IPS(R)"            "WED(R)"           
#> [163] "AL-D(R)"           "g2(R)"             "FBl5"             
#> [166] "FBl8"              "SPS(L)"            "PB(R9)"           
#> [169] "AL-VM7v(L)"        "AL-VC2(R)"         "AL-VP2(R)"        
#> [172] "AB(L)"             "RUB(R)"            "GOR(L)"           
#> [175] "mALT(R)"           "FBl3"              "AL-VP1m(R)"       
#> [178] "PB(L1)"            "AL-DA4m(L)"        "gL(R)"            
#> [181] "PB(L9)"            "bL(R)"             "AL-DC3(R)"        
#> [184] "g1(R)"             "EBr5"              "AL-DM1(R)"        
#> [187] "ICL(L)"            "ATL(R)"            "INP"              
#> [190] "EPA(R)"            "AL-DM3(L)"         "AL-DA2(L)"        
#> [193] "PB(L6)"            "PB(R1)"            "PED(R)"           
#> [196] "NO3(R)"            "AL-DM6(R)"         "AL-VM5d(R)"       
#> [199] "SMP(L)"            "ROB(R)"            "PB(L2)"           
#> [202] "CX"                "CAN(R)"            "AL-DM3(R)"        
#> [205] "VES(R)"            "FBl7"              "a'L(L)"           
#> [208] "NO2(L)"            "SPS(R)"            "SCL(R)"           
#> [211] "PB"                "AL-DC1(L)"         "a3(R)"            
#> [214] "SMP(R)"            "PB(R7)"            "LAL(-GA)(R)"      
#> [217] "SNP(L)"            "AL-VA7l(R)"        "AL-DC4(L)"        
#> [220] "AL-VM7d(L)"        "AL-DL2v(R)"        "AL-VA2(R)"        
#> [223] "AL-VM7d(R)"        "EBr3pw"            "SLP(R)"           
#> [226] "PB(R5)"            "AL-DP1l(R)"        "AL-VC3m(R)"       
#> [229] "AL-D(L)"           "AL-V(R)"          
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                        
#> [1,] "GET"  "/api/v:ver/cached/roiconnectivity"         
#> [2,] "GET"  "/api/v:ver/npexplorer/neuronmetavals"      
#> [3,] "GET"  "/api/npexplorer/autapses"                  
#> [4,] "POST" "/api/v:ver/skeletons/skeleton/:dataset/:id"
#> [5,] "POST" "/api/v:ver/npexplorer/distribution"        
#> [6,] "GET"  "/api/v:ver/cached/dailytype"               
#>      name                                                                                    
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }