Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DL4(L)"         "AL-VP2(R)"         "lACA(R)"          
#>   [4] "VMNP"              "AL-DC4(L)"         "AL-VM1(R)"        
#>   [7] "EBr3d"             "SNP(R)"            "GNG"              
#>  [10] "b'L(L)"            "AL-VM7v(R)"        "SPS(R)"           
#>  [13] "PLP(R)"            "AME(R)"            "NO(R)"            
#>  [16] "AL-VA1v(R)"        "AL-VA6(R)"         "dACA(R)"          
#>  [19] "AL-DM1(L)"         "SIP(R)"            "ME(R)"            
#>  [22] "AL-DA2(L)"         "ICL(R)"            "PB(L4)"           
#>  [25] "AL-DA2(R)"         "AL-DA4l(R)"        "AL-DL4(R)"        
#>  [28] "FB"                "AL-DA3(L)"         "NO"               
#>  [31] "VLNP(R)"           "g1(R)"             "OL(R)"            
#>  [34] "NO2(L)"            "PB(L3)"            "b'1(R)"           
#>  [37] "AL-VM2(R)"         "PB(L6)"            "ICL(L)"           
#>  [40] "PB(R4)"            "gL(R)"             "AL-DL2v(R)"       
#>  [43] "AMMC"              "AVLP(R)"           "AL-DM6(R)"        
#>  [46] "IB112"             "EBr5"              "LX(R)"            
#>  [49] "AL-DC2(L)"         "BU(L)"             "NO1(R)"           
#>  [52] "AL-VM5d(R)"        "AL-DP1l(R)"        "NO(L)"            
#>  [55] "EBr3am"            "AL-DM2(L)"         "AL-DM4(L)"        
#>  [58] "EPA(L)"            "CRE(L)"            "LO(R)"            
#>  [61] "b1(R)"             "AL-DC1(R)"         "AL-VA7m(R)"       
#>  [64] "SIP(L)"            "BU(R)"             "CRE(-RUB)(L)"     
#>  [67] "AL(R)"             "SCL(L)"            "AL-DM6(L)"        
#>  [70] "PB(R2)"            "CAN(R)"            "PB(L8)"           
#>  [73] "AL-DC3(R)"         "RUB(R)"            "AL-VP4(R)"        
#>  [76] "IB"                "b'L(R)"            "AL-DC2(R)"        
#>  [79] "AL-VA7l(R)"        "AL-VC3m(R)"        "FLA(R)"           
#>  [82] "AL-VA6(L)"         "MB(R)"             "FBl2"             
#>  [85] "LAL(R)"            "EPA(R)"            "AL(L)"            
#>  [88] "AL-DA1(R)"         "AL-DM4(R)"         "CRE(R)"           
#>  [91] "POC"               "CA(L)"             "PB(R7)"           
#>  [94] "AL-VM7d(R)"        "CX"                "PB(L9)"           
#>  [97] "a'3(R)"            "PENP"              "LAL(L)"           
#> [100] "AL-DC4(R)"         "AL-DM2(R)"         "AL-DP1m(R)"       
#> [103] "AL-VL2p(R)"        "hemibrain"         "a'2(R)"           
#> [106] "FBl5"              "SAD"               "AL-DL1(R)"        
#> [109] "AL-VP1l(R)"        "PB"                "MB(+ACA)(R)"      
#> [112] "AL-VC1(R)"         "AL-VL2a(R)"        "AL-VP3(R)"        
#> [115] "GOR(L)"            "PVLP(R)"           "AL-VA4(R)"        
#> [118] "NO2(R)"            "GOR(R)"            "LAL(-GA)(R)"      
#> [121] "AL-DL5(L)"         "ATL(R)"            "AL-VC3l(R)"       
#> [124] "AL-VP1m(R)"        "WED(R)"            "MB(L)"            
#> [127] "SNP(L)"            "LH(R)"             "SLP(R)"           
#> [130] "AL-D(R)"           "AL-VC2(R)"         "PB(L1)"           
#> [133] "AL-DM5(L)"         "EBr3pw"            "AL-VM3(R)"        
#> [136] "LOP(R)"            "AOTU(R)"           "AL-D(L)"          
#> [139] "AB(R)"             "aL(L)"             "PB(L5)"           
#> [142] "AL-DC1(L)"         "a'L(R)"            "PB(R3)"           
#> [145] "PRW"               "aL(R)"             "g2(R)"            
#> [148] "AL-DP1m(L)"        "GC"                "PB(R8)"           
#> [151] "g4(R)"             "FBl1"              "FBl6"             
#> [154] "mALT(L)"           "NO3(L)"            "AL-VP1d(R)"       
#> [157] "PB(R6)"            "a'1(R)"            "AL-DM5(R)"        
#> [160] "EBr6"              "AL-VC4(R)"         "AL-VP5(R)"        
#> [163] "RUB(L)"            "AL-VL1(R)"         "SPS(L)"           
#> [166] "AL-VM7d(L)"        "SAD(-AMMC)"        "ROB(R)"           
#> [169] "AL-DA3(R)"         "AL-V(R)"           "PB(R9)"           
#> [172] "FBl9"              "CA(R)"             "AL-VA1d(R)"       
#> [175] "AL-VA5(R)"         "AL-DL3(R)"         "AL-DL5(R)"        
#> [178] "AL-DM3(R)"         "AL-DM3(L)"         "FBl8"             
#> [181] "b'2(R)"            "a2(R)"             "EB"               
#> [184] "bL(L)"             "FB-column3"        "VES(L)"           
#> [187] "a1(R)"             "NO1(L)"            "EBr1"             
#> [190] "gL(L)"             "PED(R)"            "AL-DM1(R)"        
#> [193] "AL-VC5(R)"         "g3(R)"             "vACA(R)"          
#> [196] "CRE(-ROB,-RUB)(R)" "AL-DA4m(L)"        "AL-VM7v(L)"       
#> [199] "FBl3"              "FBl4"              "a'L(L)"           
#> [202] "AL-DL2d(R)"        "NO3(R)"            "PB(R1)"           
#> [205] "VES(R)"            "AOT(R)"            "AL-VA2(R)"        
#> [208] "GF(R)"             "mALT(R)"           "g5(R)"            
#> [211] "AB(L)"             "SMP(L)"            "FBl7"             
#> [214] "AL-VM4(R)"         "ATL(L)"            "b2(R)"            
#> [217] "SCL(R)"            "INP"               "bL(R)"            
#> [220] "AL-DA4m(R)"        "AL-VM5v(R)"        "AL-VA3(R)"        
#> [223] "EBr2r4"            "IPS(R)"            "a3(R)"            
#> [226] "PB(R5)"            "SMP(R)"            "PB(L2)"           
#> [229] "GA(R)"             "PB(L7)"            "LX(L)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "MB(L)"             "EBr2r4"            "a'1(R)"           
#>   [4] "FBl8"              "IPS(R)"            "SIP(L)"           
#>   [7] "NO2(R)"            "aL(L)"             "a'L(L)"           
#>  [10] "SNP(R)"            "a3(R)"             "FBl3"             
#>  [13] "EPA(R)"            "CRE(-ROB,-RUB)(R)" "EBr3d"            
#>  [16] "dACA(R)"           "ATL(R)"            "CX"               
#>  [19] "LAL(-GA)(R)"       "PB"                "VMNP"             
#>  [22] "MB(+ACA)(R)"       "AVLP(R)"           "vACA(R)"          
#>  [25] "aL(R)"             "PB(L2)"            "SCL(L)"           
#>  [28] "FLA(R)"            "PB(R8)"            "PB(L7)"           
#>  [31] "EBr1"              "LX(R)"             "b'L(L)"           
#>  [34] "PED(R)"            "SNP(L)"            "VES(R)"           
#>  [37] "FBl6"              "SLP(R)"            "AMMC"             
#>  [40] "SPS(L)"            "EBr6"              "PLP(R)"           
#>  [43] "ROB(R)"            "PB(R2)"            "b2(R)"            
#>  [46] "g5(R)"             "ICL(L)"            "PB(L5)"           
#>  [49] "NO2(L)"            "PENP"              "GOR(L)"           
#>  [52] "a'2(R)"            "PB(R9)"            "INP"              
#>  [55] "PB(R4)"            "GF(R)"             "NO1(R)"           
#>  [58] "ICL(R)"            "RUB(R)"            "g3(R)"            
#>  [61] "lACA(R)"           "b'L(R)"            "FBl1"             
#>  [64] "FBl7"              "VES(L)"            "a'3(R)"           
#>  [67] "mALT(L)"           "SPS(R)"            "RUB(L)"           
#>  [70] "GOR(R)"            "LAL(R)"            "POC"              
#>  [73] "LOP(R)"            "ATL(L)"            "LX(L)"            
#>  [76] "FBl5"              "CA(R)"             "AME(R)"           
#>  [79] "CAN(R)"            "MB(R)"             "PB(R7)"           
#>  [82] "PB(L6)"            "PB(L4)"            "OL(R)"            
#>  [85] "SMP(R)"            "NO1(L)"            "a2(R)"            
#>  [88] "NO"                "bL(L)"             "EBr3am"           
#>  [91] "PB(L3)"            "g1(R)"             "FB"               
#>  [94] "ME(R)"             "AL(L)"             "gL(R)"            
#>  [97] "PB(L9)"            "LH(R)"             "WED(R)"           
#> [100] "NO3(L)"            "PVLP(R)"           "EPA(L)"           
#> [103] "PB(R5)"            "CA(L)"             "CRE(L)"           
#> [106] "PB(L1)"            "PB(L8)"            "FBl2"             
#> [109] "AB(R)"             "IB"                "mALT(R)"          
#> [112] "EBr5"              "g4(R)"             "SIP(R)"           
#> [115] "b'1(R)"            "g2(R)"             "GA(R)"            
#> [118] "GC"                "BU(L)"             "EB"               
#> [121] "a1(R)"             "PB(R6)"            "VLNP(R)"          
#> [124] "LO(R)"             "gL(L)"             "BU(R)"            
#> [127] "PB(R1)"            "b1(R)"             "AL(R)"            
#> [130] "CRE(R)"            "SAD"               "a'L(R)"           
#> [133] "FBl9"              "NO(L)"             "AB(L)"            
#> [136] "SMP(L)"            "FBl4"              "b'2(R)"           
#> [139] "AOTU(R)"           "NO(R)"             "EBr3pw"           
#> [142] "PB(R3)"            "LAL(L)"            "PRW"              
#> [145] "NO3(R)"            "SCL(R)"            "SAD(-AMMC)"       
#> [148] "GNG"               "bL(R)"             "AOT(R)"           
#> [151] "hemibrain"        
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "VMNP"              "ICL(L)"            "AL-VM3(R)"        
#>   [4] "BU(R)"             "AL-VP2(R)"         "GA(R)"            
#>   [7] "ATL(L)"            "AMMC"              "b1(R)"            
#>  [10] "NO1(L)"            "EBr6"              "FB-column3"       
#>  [13] "FBl8"              "SPS(R)"            "AL-VM7v(L)"       
#>  [16] "mALT(L)"           "hemibrain"         "gL(R)"            
#>  [19] "bL(R)"             "VES(L)"            "AL-VM4(R)"        
#>  [22] "AL-VM7v(R)"        "a2(R)"             "b2(R)"            
#>  [25] "FBl9"              "OL(R)"             "LAL(L)"           
#>  [28] "AL-VP5(R)"         "VLNP(R)"           "NO2(L)"           
#>  [31] "FB"                "EBr1"              "LO(R)"            
#>  [34] "AL-DA1(R)"         "AL-VC3l(R)"        "aL(R)"            
#>  [37] "dACA(R)"           "AVLP(R)"           "LAL(R)"           
#>  [40] "AL-DM3(R)"         "AL-VA1d(R)"        "AL-VA7l(R)"       
#>  [43] "EBr5"              "NO3(L)"            "ICL(R)"           
#>  [46] "AL-DA4l(R)"        "AL-DC1(L)"         "NO"               
#>  [49] "AL-VA7m(R)"        "gL(L)"             "AL-VM5d(R)"       
#>  [52] "b'1(R)"            "g5(R)"             "FBl1"             
#>  [55] "AL-DL5(R)"         "FLA(R)"            "SNP(R)"           
#>  [58] "AL-DC2(L)"         "EBr3pw"            "GF(R)"            
#>  [61] "SCL(R)"            "AL-DM4(L)"         "MB(+ACA)(R)"      
#>  [64] "LH(R)"             "SLP(R)"            "AL-VL1(R)"        
#>  [67] "AL-VM2(R)"         "PB(R8)"            "SAD(-AMMC)"       
#>  [70] "AB(R)"             "GOR(R)"            "AL-DL3(R)"        
#>  [73] "AL-VA2(R)"         "MB(L)"             "PLP(R)"           
#>  [76] "SIP(R)"            "NO(R)"             "a'L(R)"           
#>  [79] "PB(R2)"            "AL-DM4(R)"         "AL-VM5v(R)"       
#>  [82] "AL-VA5(R)"         "SIP(L)"            "a'1(R)"           
#>  [85] "SNP(L)"            "AL-DM1(R)"         "EPA(R)"           
#>  [88] "AL-DM2(R)"         "AL-VC1(R)"         "RUB(R)"           
#>  [91] "AL-VA3(R)"         "vACA(R)"           "PB(R6)"           
#>  [94] "PB(L3)"            "PRW"               "AL-DP1m(L)"       
#>  [97] "PVLP(R)"           "AL-DC2(R)"         "bL(L)"            
#> [100] "AL-DC3(R)"         "AL-DA3(L)"         "AL-DM5(R)"        
#> [103] "AOT(R)"            "AL-DA2(R)"         "AL-VL2a(R)"       
#> [106] "AL-VM7d(R)"        "AL-V(R)"           "g3(R)"            
#> [109] "LAL(-GA)(R)"       "AL-DL4(L)"         "AL-VM7d(L)"       
#> [112] "CRE(R)"            "PB(L9)"            "NO2(R)"           
#> [115] "AL-DP1m(R)"        "PED(R)"            "NO(L)"            
#> [118] "a'L(L)"            "AL-VP4(R)"         "SAD"              
#> [121] "PB"                "AME(R)"            "AL-VC4(R)"        
#> [124] "ME(R)"             "g4(R)"             "ATL(R)"           
#> [127] "LX(L)"             "AL-DM6(L)"         "PB(L1)"           
#> [130] "FBl5"              "AB(L)"             "AL(L)"            
#> [133] "EPA(L)"            "POC"               "PB(L2)"           
#> [136] "AL-VC3m(R)"        "AL-VM1(R)"         "LOP(R)"           
#> [139] "SMP(L)"            "AL-VA6(L)"         "FBl3"             
#> [142] "PB(R3)"            "AL-DA2(L)"         "FBl6"             
#> [145] "CA(R)"             "AL-DA4m(R)"        "AL-DM1(L)"        
#> [148] "FBl4"              "a3(R)"             "MB(R)"            
#> [151] "AL-DM6(R)"         "AL-VA6(R)"         "GC"               
#> [154] "a'2(R)"            "PB(R5)"            "VES(R)"           
#> [157] "EBr3am"            "a'3(R)"            "PB(L8)"           
#> [160] "PB(R9)"            "AL(R)"             "AL-VL2p(R)"       
#> [163] "AL-DM3(L)"         "AL-DC4(R)"         "AL-DM2(L)"        
#> [166] "RUB(L)"            "AL-VA4(R)"         "AL-VP1d(R)"       
#> [169] "FBl2"              "AL-DL1(R)"         "AL-VA1v(R)"       
#> [172] "LX(R)"             "AL-DA4m(L)"        "CRE(L)"           
#> [175] "PB(L4)"            "AL-DA3(R)"         "CX"               
#> [178] "AL-DM5(L)"         "PB(L6)"            "CRE(-RUB)(L)"     
#> [181] "CAN(R)"            "PENP"              "GOR(L)"           
#> [184] "NO1(R)"            "CRE(-ROB,-RUB)(R)" "ROB(R)"           
#> [187] "b'L(L)"            "SPS(L)"            "AL-D(R)"          
#> [190] "AL-VP1m(R)"        "PB(R7)"            "mALT(R)"          
#> [193] "WED(R)"            "b'L(R)"            "AL-DL5(L)"        
#> [196] "AL-VP1l(R)"        "AL-D(L)"           "AL-DL2v(R)"       
#> [199] "AL-VC5(R)"         "g1(R)"             "AL-VP3(R)"        
#> [202] "IB"                "AOTU(R)"           "AL-DC4(L)"        
#> [205] "FBl7"              "a1(R)"             "NO3(R)"           
#> [208] "EB"                "PB(R1)"            "IPS(R)"           
#> [211] "AL-VC2(R)"         "CA(L)"             "b'2(R)"           
#> [214] "EBr2r4"            "PB(R4)"            "AL-DL2d(R)"       
#> [217] "AL-DL4(R)"         "PB(L7)"            "GNG"              
#> [220] "SCL(L)"            "SMP(R)"            "PB(L5)"           
#> [223] "EBr3d"             "aL(L)"             "INP"              
#> [226] "AL-DC1(R)"         "lACA(R)"           "g2(R)"            
#> [229] "BU(L)"             "AL-DP1l(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "gL(L)"             "AL-DM1(L)"         "SNP(L)"           
#>   [4] "MB(+ACA)(R)"       "LX(L)"             "AL-VP5(R)"        
#>   [7] "a'3(R)"            "PB(L3)"            "AL-DC1(L)"        
#>  [10] "AL-DP1m(L)"        "NO3(L)"            "AB(R)"            
#>  [13] "AL-VL2p(R)"        "g4(R)"             "bL(R)"            
#>  [16] "AL-VA4(R)"         "g3(R)"             "AL-VM7v(L)"       
#>  [19] "AL-DM2(R)"         "AL-VC5(R)"         "PB(R6)"           
#>  [22] "INP"               "AL-DL5(L)"         "AL-DM6(L)"        
#>  [25] "FBl6"              "AL-DA1(R)"         "AL-VP2(R)"        
#>  [28] "CRE(-ROB,-RUB)(R)" "LAL(L)"            "EBr3d"            
#>  [31] "AME(R)"            "GNG"               "PRW"              
#>  [34] "VES(R)"            "AL-DC1(R)"         "PB(R8)"           
#>  [37] "ICL(L)"            "CA(R)"             "b1(R)"            
#>  [40] "AL-DL1(R)"         "AL-DA3(L)"         "AL-DM3(L)"        
#>  [43] "OL(R)"             "AL-DL3(R)"         "AL-VM4(R)"        
#>  [46] "AVLP(R)"           "LAL(R)"            "g5(R)"            
#>  [49] "FBl8"              "NO(L)"             "AL-VA1d(R)"       
#>  [52] "AL-VC2(R)"         "AL-VM7d(R)"        "ATL(R)"           
#>  [55] "AL-VP3(R)"         "PB(L2)"            "PB(L4)"           
#>  [58] "GF(R)"             "lACA(R)"           "PB(L9)"           
#>  [61] "PB(L7)"            "FBl3"              "PB(R3)"           
#>  [64] "AL-VA5(R)"         "AL-VA7m(R)"        "VLNP(R)"          
#>  [67] "SIP(L)"            "a'1(R)"            "AL-VC3l(R)"       
#>  [70] "PVLP(R)"           "b'L(L)"            "GC"               
#>  [73] "AL-VC3m(R)"        "SPS(L)"            "AL-DA4m(L)"       
#>  [76] "PB(L6)"            "IPS(R)"            "AL-VP4(R)"        
#>  [79] "PLP(R)"            "AL-VM3(R)"         "AL-DC2(L)"        
#>  [82] "FBl4"              "LH(R)"             "AL-VA7l(R)"       
#>  [85] "CX"                "FBl1"              "FBl5"             
#>  [88] "AOT(R)"            "NO1(L)"            "EBr5"             
#>  [91] "EBr3am"            "dACA(R)"           "AL-DL4(L)"        
#>  [94] "AL-VC4(R)"         "WED(R)"            "AL-DM2(L)"        
#>  [97] "AL-VP1l(R)"        "GOR(R)"            "AL-V(R)"          
#> [100] "AL-VP1d(R)"        "FBl9"              "SCL(L)"           
#> [103] "hemibrain"         "PB(R7)"            "PED(R)"           
#> [106] "PB"                "AL-DL2d(R)"        "aL(R)"            
#> [109] "AL-VA6(L)"         "NO"                "AL-DA2(L)"        
#> [112] "AL-VM2(R)"         "PB(R9)"            "NO2(R)"           
#> [115] "AL-VL1(R)"         "mALT(R)"           "AL-DA4l(R)"       
#> [118] "AL-DC3(R)"         "gL(R)"             "ATL(L)"           
#> [121] "SMP(L)"            "CRE(R)"            "VMNP"             
#> [124] "ICL(R)"            "AL-DL2v(R)"        "AL-VA2(R)"        
#> [127] "AL-VM7d(L)"        "AL(R)"             "AL(L)"            
#> [130] "AL-DC4(L)"         "AL-VM5v(R)"        "AL-VM7v(R)"       
#> [133] "NO2(L)"            "a2(R)"             "PB(R1)"           
#> [136] "g1(R)"             "a3(R)"             "AL-VA3(R)"        
#> [139] "CAN(R)"            "aL(L)"             "b'1(R)"           
#> [142] "CA(L)"             "NO3(R)"            "ME(R)"            
#> [145] "AL-DM5(L)"         "IB"                "AMMC"             
#> [148] "EB"                "AL-DA3(R)"         "AL-DP1m(R)"       
#> [151] "AOTU(R)"           "SIP(R)"            "bL(L)"            
#> [154] "PB(R4)"            "NO1(R)"            "EPA(R)"           
#> [157] "AL-DM3(R)"         "ROB(R)"            "EBr3pw"           
#> [160] "AL-DL4(R)"         "AL-VA6(R)"         "SLP(R)"           
#> [163] "AL-VL2a(R)"        "mALT(L)"           "AL-DM4(L)"        
#> [166] "AL-DA2(R)"         "RUB(R)"            "GA(R)"            
#> [169] "SPS(R)"            "GOR(L)"            "POC"              
#> [172] "SAD"               "LO(R)"             "FBl2"             
#> [175] "AL-VA1v(R)"        "LOP(R)"            "b'2(R)"           
#> [178] "FB"                "AL-D(R)"           "AL-VP1m(R)"       
#> [181] "AL-DA4m(R)"        "AL-DC4(R)"         "a'L(R)"           
#> [184] "SAD(-AMMC)"        "FB-column3"        "AL-DC2(R)"        
#> [187] "PENP"              "LAL(-GA)(R)"       "BU(R)"            
#> [190] "PB(R2)"            "SCL(R)"            "FBl7"             
#> [193] "a'L(L)"            "a1(R)"             "MB(L)"            
#> [196] "CRE(L)"            "PB(L1)"            "SNP(R)"           
#> [199] "AL-D(L)"           "CRE(-RUB)(L)"      "AL-VM5d(R)"       
#> [202] "PB(L5)"            "b2(R)"             "AL-DM5(R)"        
#> [205] "AL-DM6(R)"         "PB(L8)"            "EPA(L)"           
#> [208] "RUB(L)"            "AB(L)"             "FLA(R)"           
#> [211] "NO(R)"             "VES(L)"            "AL-DL5(R)"        
#> [214] "vACA(R)"           "g2(R)"             "LX(R)"            
#> [217] "a'2(R)"            "MB(R)"             "AL-DM1(R)"        
#> [220] "PB(R5)"            "b'L(R)"            "EBr2r4"           
#> [223] "AL-VM1(R)"         "SMP(R)"            "EBr1"             
#> [226] "EBr6"              "BU(L)"             "AL-DM4(R)"        
#> [229] "AL-DP1l(R)"        "AL-VC1(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                                  
#> [1,] "CONNECT" "/api"                                
#> [2,] "GET"     "/api/v:ver/npexplorer/findneurons"   
#> [3,] "GET"     "/api/raw/keyvalue/key/:instance/:key"
#> [4,] "GET"     "/api/v:ver/custom/custom"            
#> [5,] "POST"    "/api/custom/custom"                  
#> [6,] "POST"    "/api/v:ver/custom/custom"            
#>      name                                                                             
#> [1,] "github.com/labstack/echo.glob..func1"                                           
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"             
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/raw/keyvalue.masterAPI.getKV-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"             
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"  
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"             

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }