Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "SPS(R)"            "bL(R)"             "AL-VM7v(R)"       
#>   [4] "mALT(L)"           "PB(R8)"            "AL-DL5(L)"        
#>   [7] "FBl3"              "AL-DA2(L)"         "AL-DC2(L)"        
#>  [10] "AL-DL5(R)"         "AL(R)"             "AL-VP4(R)"        
#>  [13] "AL-DM3(L)"         "IB112"             "VES(R)"           
#>  [16] "AVLP(R)"           "AL-DM2(R)"         "AOT(R)"           
#>  [19] "SAD"               "NO1(R)"            "IPS(R)"           
#>  [22] "NO(R)"             "ICL(L)"            "PB(R1)"           
#>  [25] "FBl7"              "AL-DA4l(R)"        "AL-DA2(R)"        
#>  [28] "CRE(R)"            "VMNP"              "PB(L6)"           
#>  [31] "LAL(R)"            "SLP(R)"            "lACA(R)"          
#>  [34] "NO3(R)"            "gL(R)"             "LO(R)"            
#>  [37] "AL-VM7v(L)"        "CAN(R)"            "AL-VA6(R)"        
#>  [40] "AL-DA1(R)"         "AL-VA3(R)"         "AL-VC2(R)"        
#>  [43] "POC"               "AL-VM5d(R)"        "AL-V(R)"          
#>  [46] "AL-DM6(L)"         "FBl5"              "AL-VA2(R)"        
#>  [49] "AL-VA7m(R)"        "AL-VC3l(R)"        "AL-VM7d(R)"       
#>  [52] "PB(L2)"            "EBr3d"             "SIP(R)"           
#>  [55] "CRE(L)"            "FBl1"              "MB(L)"            
#>  [58] "AL-VL2a(R)"        "CX"                "g3(R)"            
#>  [61] "ATL(R)"            "AL-DL3(R)"         "AL-VM1(R)"        
#>  [64] "AL-DC4(L)"         "EBr3pw"            "AL-VM3(R)"        
#>  [67] "SNP(L)"            "EBr3am"            "WED(R)"           
#>  [70] "PB(L1)"            "a'L(L)"            "AL-D(R)"          
#>  [73] "AL-VM4(R)"         "PRW"               "NO1(L)"           
#>  [76] "AL-DM4(L)"         "AL-DP1m(L)"        "GOR(L)"           
#>  [79] "PB(R9)"            "PB(R3)"            "ROB(R)"           
#>  [82] "AL-DA3(R)"         "a1(R)"             "SPS(L)"           
#>  [85] "NO"                "AB(R)"             "AL-DC3(R)"        
#>  [88] "PLP(R)"            "AL-DM2(L)"         "b1(R)"            
#>  [91] "AL-VA1d(R)"        "AL-DC2(R)"         "AMMC"             
#>  [94] "FB"                "AL-DL1(R)"         "AL-DL2v(R)"       
#>  [97] "AL-DP1m(R)"        "a'L(R)"            "PB(R4)"           
#> [100] "AL-DC4(R)"         "CA(L)"             "PB(L4)"           
#> [103] "PB(L5)"            "LAL(L)"            "g2(R)"            
#> [106] "EB"                "a3(R)"             "NO2(L)"           
#> [109] "ME(R)"             "PB(L9)"            "PB(L7)"           
#> [112] "AL-VC1(R)"         "GC"                "AME(R)"           
#> [115] "EBr1"              "PB(L8)"            "MB(+ACA)(R)"      
#> [118] "OL(R)"             "AL-VM2(R)"         "AL-VP5(R)"        
#> [121] "b'1(R)"            "b2(R)"             "AL(L)"            
#> [124] "FBl6"              "AL-DL2d(R)"        "EPA(L)"           
#> [127] "AL-DP1l(R)"        "AL-VL2p(R)"        "bL(L)"            
#> [130] "GOR(R)"            "BU(R)"             "FBl9"             
#> [133] "AL-DM6(R)"         "PB(R2)"            "AL-VA7l(R)"       
#> [136] "AL-VC4(R)"         "SMP(L)"            "aL(R)"            
#> [139] "FBl2"              "AL-VP1d(R)"        "a'3(R)"           
#> [142] "a'1(R)"            "FBl8"              "NO2(R)"           
#> [145] "VES(L)"            "AL-VA6(L)"         "AL-VM7d(L)"       
#> [148] "aL(L)"             "b'L(L)"            "LAL(-GA)(R)"      
#> [151] "FLA(R)"            "LX(R)"             "a'2(R)"           
#> [154] "AL-VA1v(R)"        "IB"                "AOTU(R)"          
#> [157] "AL-DL4(L)"         "g1(R)"             "gL(L)"            
#> [160] "AL-DM5(L)"         "AL-VM5v(R)"        "MB(R)"            
#> [163] "b'2(R)"            "GA(R)"             "b'L(R)"           
#> [166] "INP"               "hemibrain"         "LX(L)"            
#> [169] "NO(L)"             "AL-VA5(R)"         "PED(R)"           
#> [172] "SNP(R)"            "AL-DL4(R)"         "AL-VA4(R)"        
#> [175] "a2(R)"             "EBr2r4"            "AL-DC1(L)"        
#> [178] "AL-DM3(R)"         "SIP(L)"            "AL-DM5(R)"        
#> [181] "AL-VP1l(R)"        "PB(R7)"            "AL-VP3(R)"        
#> [184] "AL-DA3(L)"         "AL-VL1(R)"         "mALT(R)"          
#> [187] "g4(R)"             "FBl4"              "ICL(R)"           
#> [190] "AL-VP2(R)"         "AB(L)"             "EBr6"             
#> [193] "dACA(R)"           "GNG"               "AL-DA4m(R)"       
#> [196] "LOP(R)"            "PB(R6)"            "EBr5"             
#> [199] "SAD(-AMMC)"        "FB-column3"        "g5(R)"            
#> [202] "RUB(L)"            "AL-VC5(R)"         "RUB(R)"           
#> [205] "EPA(R)"            "CRE(-RUB)(L)"      "VLNP(R)"          
#> [208] "SCL(L)"            "PB(L3)"            "NO3(L)"           
#> [211] "AL-DM1(R)"         "SMP(R)"            "vACA(R)"          
#> [214] "PB"                "CA(R)"             "AL-VC3m(R)"       
#> [217] "AL-VP1m(R)"        "CRE(-ROB,-RUB)(R)" "PENP"             
#> [220] "BU(L)"             "AL-DM1(L)"         "PVLP(R)"          
#> [223] "SCL(R)"            "AL-DC1(R)"         "AL-DM4(R)"        
#> [226] "ATL(L)"            "GF(R)"             "AL-D(L)"          
#> [229] "LH(R)"             "PB(R5)"            "AL-DA4m(L)"       
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PED(R)"            "PB(R4)"            "PB(R5)"           
#>   [4] "a3(R)"             "AL(R)"             "POC"              
#>   [7] "AMMC"              "aL(R)"             "g5(R)"            
#>  [10] "ICL(R)"            "LX(L)"             "RUB(L)"           
#>  [13] "VES(L)"            "SIP(L)"            "EBr2r4"           
#>  [16] "BU(R)"             "LAL(-GA)(R)"       "AOT(R)"           
#>  [19] "RUB(R)"            "b1(R)"             "IPS(R)"           
#>  [22] "SCL(L)"            "CRE(L)"            "PB(L7)"           
#>  [25] "FBl7"              "ATL(R)"            "LAL(R)"           
#>  [28] "NO1(R)"            "a'L(R)"            "PB(R1)"           
#>  [31] "a'1(R)"            "a2(R)"             "dACA(R)"          
#>  [34] "PB(L1)"            "PB(R8)"            "EB"               
#>  [37] "gL(L)"             "AVLP(R)"           "IB"               
#>  [40] "AOTU(R)"           "EBr5"              "BU(L)"            
#>  [43] "EBr3d"             "NO(R)"             "AB(L)"            
#>  [46] "aL(L)"             "CRE(-ROB,-RUB)(R)" "PLP(R)"           
#>  [49] "ICL(L)"            "FBl8"              "VMNP"             
#>  [52] "PB(R9)"            "ROB(R)"            "LH(R)"            
#>  [55] "GA(R)"             "NO1(L)"            "SCL(R)"           
#>  [58] "EBr3am"            "a1(R)"             "CA(L)"            
#>  [61] "FBl3"              "OL(R)"             "MB(R)"            
#>  [64] "b'L(R)"            "EBr1"              "PB(L2)"           
#>  [67] "vACA(R)"           "CX"                "PENP"             
#>  [70] "GOR(L)"            "MB(L)"             "GOR(R)"           
#>  [73] "NO(L)"             "LOP(R)"            "PB(R6)"           
#>  [76] "EBr3pw"            "PB(R2)"            "SIP(R)"           
#>  [79] "PB(L6)"            "FBl4"              "INP"              
#>  [82] "a'3(R)"            "CRE(R)"            "EPA(R)"           
#>  [85] "SNP(L)"            "PB(L8)"            "AME(R)"           
#>  [88] "a'L(L)"            "SMP(L)"            "g3(R)"            
#>  [91] "gL(R)"             "g2(R)"             "EPA(L)"           
#>  [94] "SLP(R)"            "PB(R3)"            "AB(R)"            
#>  [97] "b'1(R)"            "SPS(R)"            "mALT(R)"          
#> [100] "bL(L)"             "CA(R)"             "PB(L4)"           
#> [103] "g4(R)"             "VES(R)"            "MB(+ACA)(R)"      
#> [106] "PB(L5)"            "GF(R)"             "lACA(R)"          
#> [109] "NO"                "b2(R)"             "AL(L)"            
#> [112] "NO2(L)"            "FLA(R)"            "FBl5"             
#> [115] "GNG"               "PB(R7)"            "GC"               
#> [118] "VLNP(R)"           "a'2(R)"            "FBl6"             
#> [121] "bL(R)"             "LX(R)"             "ATL(L)"           
#> [124] "PB(L3)"            "NO3(R)"            "PB"               
#> [127] "NO3(L)"            "PVLP(R)"           "FBl1"             
#> [130] "LAL(L)"            "FB"                "ME(R)"            
#> [133] "CAN(R)"            "b'L(L)"            "WED(R)"           
#> [136] "g1(R)"             "LO(R)"             "SMP(R)"           
#> [139] "PB(L9)"            "FBl2"              "NO2(R)"           
#> [142] "EBr6"              "SNP(R)"            "PRW"              
#> [145] "SPS(L)"            "SAD(-AMMC)"        "FBl9"             
#> [148] "mALT(L)"           "b'2(R)"            "hemibrain"        
#> [151] "SAD"              
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "PENP"              "VES(R)"            "AL-VP3(R)"        
#>   [4] "g5(R)"             "EPA(L)"            "AB(L)"            
#>   [7] "SAD(-AMMC)"        "AL-VP4(R)"         "aL(R)"            
#>  [10] "gL(R)"             "CRE(-RUB)(L)"      "AL-DC4(R)"        
#>  [13] "AL-VM2(R)"         "PB(L2)"            "MB(L)"            
#>  [16] "FLA(R)"            "NO3(L)"            "NO(L)"            
#>  [19] "AOTU(R)"           "AL-DP1m(L)"        "AL-DC1(R)"        
#>  [22] "AL-DL4(R)"         "AL-VM3(R)"         "AL-VM5d(R)"       
#>  [25] "PB(R4)"            "AL-DA1(R)"         "GF(R)"            
#>  [28] "g3(R)"             "CRE(L)"            "FBl2"             
#>  [31] "AL-VA6(R)"         "BU(L)"             "AME(R)"           
#>  [34] "AL-DM3(R)"         "AL-VA1d(R)"        "PB(L3)"           
#>  [37] "AL-DM4(R)"         "PB(R6)"            "PB(R5)"           
#>  [40] "CRE(-ROB,-RUB)(R)" "NO2(L)"            "PB"               
#>  [43] "AL-DM3(L)"         "MB(R)"             "AL-VP1m(R)"       
#>  [46] "AL-DL3(R)"         "CA(L)"             "NO(R)"            
#>  [49] "ICL(L)"            "LX(L)"             "SMP(R)"           
#>  [52] "PB(L8)"            "AL-DC4(L)"         "FBl6"             
#>  [55] "AL-DA4l(R)"        "AL-VP2(R)"         "AL-VA6(L)"        
#>  [58] "LAL(R)"            "AL-DC2(L)"         "AL-VM7v(R)"       
#>  [61] "a2(R)"             "GC"                "dACA(R)"          
#>  [64] "EBr6"              "EBr2r4"            "AL-DL5(L)"        
#>  [67] "a1(R)"             "a'3(R)"            "PB(L5)"           
#>  [70] "EBr3d"             "EB"                "AL-DM6(R)"        
#>  [73] "AL-V(R)"           "CX"                "SPS(R)"           
#>  [76] "FB"                "ROB(R)"            "FBl3"             
#>  [79] "AL-VA4(R)"         "AL-VM5v(R)"        "PRW"              
#>  [82] "PB(R7)"            "AL-DM4(L)"         "CAN(R)"           
#>  [85] "AB(R)"             "AL-VL1(R)"         "hemibrain"        
#>  [88] "PB(R8)"            "EBr3pw"            "CA(R)"            
#>  [91] "AL-VP5(R)"         "LO(R)"             "PB(R2)"           
#>  [94] "OL(R)"             "AL-VC2(R)"         "EBr5"             
#>  [97] "EPA(R)"            "PB(L7)"            "SNP(L)"           
#> [100] "NO2(R)"            "PB(L4)"            "AL-VC3l(R)"       
#> [103] "POC"               "AL(R)"             "AL-VM1(R)"        
#> [106] "EBr3am"            "AL-VL2a(R)"        "AL-VM4(R)"        
#> [109] "IPS(R)"            "GOR(R)"            "AL-DC3(R)"        
#> [112] "AL-VC1(R)"         "lACA(R)"           "a'2(R)"           
#> [115] "PVLP(R)"           "FBl7"              "SCL(L)"           
#> [118] "RUB(L)"            "a'L(L)"            "AL-DM5(R)"        
#> [121] "AL-DP1m(R)"        "PB(L1)"            "AL-DL5(R)"        
#> [124] "AL-VA7l(R)"        "vACA(R)"           "GOR(L)"           
#> [127] "PB(R1)"            "AL-VC3m(R)"        "AL-DA3(L)"        
#> [130] "AL-DM2(L)"         "b'1(R)"            "AMMC"             
#> [133] "b'L(R)"            "SNP(R)"            "b1(R)"            
#> [136] "AL-VC4(R)"         "a3(R)"             "LX(R)"            
#> [139] "a'L(R)"            "AL-DC1(L)"         "FBl4"             
#> [142] "FBl5"              "AL-VA3(R)"         "mALT(R)"          
#> [145] "MB(+ACA)(R)"       "AL-DC2(R)"         "AL-D(R)"          
#> [148] "NO3(R)"            "AVLP(R)"           "AL-DL2v(R)"       
#> [151] "LAL(-GA)(R)"       "PB(L6)"            "a'1(R)"           
#> [154] "VLNP(R)"           "NO1(R)"            "ICL(R)"           
#> [157] "PLP(R)"            "g4(R)"             "INP"              
#> [160] "SIP(R)"            "PB(R9)"            "FBl8"             
#> [163] "RUB(R)"            "SMP(L)"            "WED(R)"           
#> [166] "CRE(R)"            "AL-VL2p(R)"        "AL-DM1(R)"        
#> [169] "IB"                "g1(R)"             "BU(R)"            
#> [172] "AL-DL4(L)"         "bL(R)"             "VMNP"             
#> [175] "AL-VA7m(R)"        "AL-VP1d(R)"        "mALT(L)"          
#> [178] "AL-VM7d(L)"        "g2(R)"             "AL-DA4m(L)"       
#> [181] "aL(L)"             "AL-VA5(R)"         "AL-DA3(R)"        
#> [184] "SIP(L)"            "AL-D(L)"           "NO"               
#> [187] "FB-column3"        "LH(R)"             "b'L(L)"           
#> [190] "ATL(L)"            "GA(R)"             "ME(R)"            
#> [193] "AL-DM5(L)"         "AL-VA1v(R)"        "SAD"              
#> [196] "b2(R)"             "gL(L)"             "AL-DA2(L)"        
#> [199] "FBl1"              "GNG"               "AL-DM2(R)"        
#> [202] "AL-VM7v(L)"        "bL(L)"             "AL-VM7d(R)"       
#> [205] "LOP(R)"            "PB(R3)"            "AL-VP1l(R)"       
#> [208] "AOT(R)"            "b'2(R)"            "NO1(L)"           
#> [211] "SPS(L)"            "ATL(R)"            "AL-DP1l(R)"       
#> [214] "SCL(R)"            "AL-DM1(L)"         "FBl9"             
#> [217] "AL-DA4m(R)"        "AL-VC5(R)"         "AL-DM6(L)"        
#> [220] "VES(L)"            "AL-DL1(R)"         "AL-DL2d(R)"       
#> [223] "PED(R)"            "PB(L9)"            "SLP(R)"           
#> [226] "AL-VA2(R)"         "LAL(L)"            "EBr1"             
#> [229] "AL(L)"             "AL-DA2(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-DA2(L)"         "FBl1"              "PB(L4)"           
#>   [4] "g3(R)"             "EPA(L)"            "AL-DA3(R)"        
#>   [7] "RUB(R)"            "SNP(R)"            "EBr3pw"           
#>  [10] "bL(L)"             "VLNP(R)"           "hemibrain"        
#>  [13] "ME(R)"             "LAL(R)"            "AL-DM3(R)"        
#>  [16] "b2(R)"             "PB(L8)"            "PB(R1)"           
#>  [19] "FBl2"              "ICL(R)"            "EBr3d"            
#>  [22] "AL-DM2(L)"         "AL(R)"             "EBr3am"           
#>  [25] "VES(R)"            "b'L(L)"            "EB"               
#>  [28] "IPS(R)"            "LAL(L)"            "PLP(R)"           
#>  [31] "GC"                "AL-VM5d(R)"        "AL-DL4(R)"        
#>  [34] "mALT(L)"           "AL-DM4(L)"         "AL-VL2p(R)"       
#>  [37] "AMMC"              "GA(R)"             "b'L(R)"           
#>  [40] "LH(R)"             "NO(L)"             "a'L(L)"           
#>  [43] "MB(R)"             "AL-VC3m(R)"        "AL-VM7d(R)"       
#>  [46] "g1(R)"             "gL(R)"             "POC"              
#>  [49] "AL-VA2(R)"         "SIP(L)"            "AL-VA1d(R)"       
#>  [52] "GOR(L)"            "AL-DL3(R)"         "AL(L)"            
#>  [55] "AL-DP1m(L)"        "g5(R)"             "OL(R)"            
#>  [58] "AL-DM6(R)"         "AL-DP1m(R)"        "PENP"             
#>  [61] "EBr6"              "AL-V(R)"           "AL-VM1(R)"        
#>  [64] "NO1(R)"            "AL-DC2(L)"         "GNG"              
#>  [67] "PB(L9)"            "NO3(L)"            "CRE(-ROB,-RUB)(R)"
#>  [70] "CRE(L)"            "AL-VM4(R)"         "NO2(L)"           
#>  [73] "AL-DC4(L)"         "AL-DC4(R)"         "AL-DM3(L)"        
#>  [76] "AL-VM7v(L)"        "FBl6"              "LX(L)"            
#>  [79] "CAN(R)"            "AL-DM4(R)"         "SCL(R)"           
#>  [82] "BU(L)"             "AL-VM5v(R)"        "AL-VC3l(R)"       
#>  [85] "AL-DM5(L)"         "AL-VA7l(R)"        "PB(R5)"           
#>  [88] "PED(R)"            "SNP(L)"            "AL-DM1(L)"        
#>  [91] "AL-DL1(R)"         "mALT(R)"           "NO3(R)"           
#>  [94] "EBr5"              "LX(R)"             "PB(R2)"           
#>  [97] "AL-VP1d(R)"        "SPS(R)"            "PB(R7)"           
#> [100] "ROB(R)"            "bL(R)"             "GOR(R)"           
#> [103] "PB(L1)"            "EBr2r4"            "CA(R)"            
#> [106] "AL-VP3(R)"         "AL-VA1v(R)"        "AL-VA4(R)"        
#> [109] "AOT(R)"            "SAD"               "AL-VA5(R)"        
#> [112] "AL-VC4(R)"         "b'1(R)"            "AL-VC1(R)"        
#> [115] "AL-VM3(R)"         "CA(L)"             "lACA(R)"          
#> [118] "AL-DL4(L)"         "PB(R9)"            "NO(R)"            
#> [121] "AL-DA2(R)"         "AL-VP2(R)"         "PB(L2)"           
#> [124] "GF(R)"             "PB"                "SMP(R)"           
#> [127] "a'3(R)"            "PB(L3)"            "AB(L)"            
#> [130] "CRE(-RUB)(L)"      "FBl7"              "NO1(L)"           
#> [133] "a'2(R)"            "AL-DA4m(L)"        "SLP(R)"           
#> [136] "AL-VP1l(R)"        "VMNP"              "EPA(R)"           
#> [139] "AL-DA3(L)"         "FBl4"              "ATL(L)"           
#> [142] "SMP(L)"            "AL-VM7d(L)"        "ATL(R)"           
#> [145] "PB(R8)"            "FBl8"              "AL-VC2(R)"        
#> [148] "vACA(R)"           "AL-VA6(R)"         "LAL(-GA)(R)"      
#> [151] "aL(L)"             "MB(L)"             "PB(L7)"           
#> [154] "VES(L)"            "a1(R)"             "dACA(R)"          
#> [157] "NO"                "RUB(L)"            "BU(R)"            
#> [160] "AL-VA6(L)"         "CX"                "AL-DA4m(R)"       
#> [163] "AL-DL2d(R)"        "AL-D(L)"           "AB(R)"            
#> [166] "gL(L)"             "AL-DL5(L)"         "SCL(L)"           
#> [169] "a'L(R)"            "AL-DC1(L)"         "AL-VA3(R)"        
#> [172] "AL-VL2a(R)"        "b'2(R)"            "PVLP(R)"          
#> [175] "AL-VC5(R)"         "MB(+ACA)(R)"       "AL-DA4l(R)"       
#> [178] "AL-DC1(R)"         "AL-VP4(R)"         "aL(R)"            
#> [181] "FBl3"              "AL-VP1m(R)"        "LOP(R)"           
#> [184] "PB(L6)"            "b1(R)"             "AME(R)"           
#> [187] "AL-DM1(R)"         "AL-DM5(R)"         "AL-VP5(R)"        
#> [190] "AOTU(R)"           "ICL(L)"            "SIP(R)"           
#> [193] "AL-DA1(R)"         "PB(R3)"            "PB(R4)"           
#> [196] "PB(L5)"            "PB(R6)"            "SPS(L)"           
#> [199] "a'1(R)"            "AL-DL5(R)"         "CRE(R)"           
#> [202] "FB"                "AL-VL1(R)"         "a2(R)"            
#> [205] "g4(R)"             "a3(R)"             "AL-D(R)"          
#> [208] "AL-VM2(R)"         "SAD(-AMMC)"        "AL-DC3(R)"        
#> [211] "NO2(R)"            "AL-VA7m(R)"        "FBl5"             
#> [214] "AL-DL2v(R)"        "AL-DP1l(R)"        "AL-VM7v(R)"       
#> [217] "FLA(R)"            "FB-column3"        "AL-DC2(R)"        
#> [220] "IB"                "g2(R)"             "LO(R)"            
#> [223] "AL-DM6(L)"         "FBl9"              "AVLP(R)"          
#> [226] "PRW"               "WED(R)"            "EBr1"             
#> [229] "INP"               "AL-DM2(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                           
#> [1,] "POST" "/api/v:ver/raw/cypher/transaction/:id/cypher" 
#> [2,] "GET"  "/api/v:ver/npexplorer/roisinneuron"           
#> [3,] "GET"  "/api/v:ver/npexplorer/commonconnectivity"     
#> [4,] "GET"  "/api/v:ver/npexplorer/celltype/:dataset/:type"
#> [5,] "GET"  "/profile"                                     
#> [6,] "POST" "/api/npexplorer/simpleconnections"            
#>      name                                                                                             
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [5,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                         
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }