Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "SCL(L)"            "PB(L8)"            "AL-DC2(L)"        
#>   [4] "PB(R9)"            "AVLP(R)"           "LAL(R)"           
#>   [7] "AL-VA6(R)"         "AL-VC3l(R)"        "ROB(R)"           
#>  [10] "OL(R)"             "AL-VP4(R)"         "PLP(R)"           
#>  [13] "a3(R)"             "AL-VA7l(R)"        "AL-VC4(R)"        
#>  [16] "PRW"               "PB(R7)"            "EPA(R)"           
#>  [19] "FBl4"              "PB(R3)"            "INP"              
#>  [22] "AL-VC2(R)"         "AL-VL2a(R)"        "AL-DM4(L)"        
#>  [25] "FBl9"              "NO2(L)"            "PENP"             
#>  [28] "PB(R8)"            "AL-VA6(L)"         "PVLP(R)"          
#>  [31] "AL-DM6(R)"         "LOP(R)"            "AOTU(R)"          
#>  [34] "SCL(R)"            "NO(R)"             "AL-DC3(R)"        
#>  [37] "AL-VM5d(R)"        "AL-VM7v(R)"        "AL-VP5(R)"        
#>  [40] "SAD"               "CRE(R)"            "FBl5"             
#>  [43] "a'L(L)"            "SLP(R)"            "PB(R4)"           
#>  [46] "GA(R)"             "PED(R)"            "PB"               
#>  [49] "PB(L2)"            "AL-VP3(R)"         "LX(L)"            
#>  [52] "RUB(L)"            "AL-VA5(R)"         "mALT(L)"          
#>  [55] "gL(L)"             "AL-DM3(L)"         "NO1(R)"           
#>  [58] "EBr2r4"            "ICL(L)"            "IPS(R)"           
#>  [61] "AL-DA2(R)"         "g3(R)"             "GC"               
#>  [64] "EBr1"              "bL(L)"             "AL-DA1(R)"        
#>  [67] "AL-VA3(R)"         "SPS(L)"            "FB-column3"       
#>  [70] "aL(L)"             "SNP(L)"            "AL-DC2(R)"        
#>  [73] "GOR(L)"            "g4(R)"             "AL-DA3(L)"        
#>  [76] "AME(R)"            "IB"                "a'2(R)"           
#>  [79] "AL-DM5(L)"         "LH(R)"             "EPA(L)"           
#>  [82] "AL-DL2d(R)"        "AL-DM6(L)"         "AL-VA7m(R)"       
#>  [85] "AL-DA2(L)"         "SIP(R)"            "ATL(R)"           
#>  [88] "AL-VA1v(R)"        "AL-VM1(R)"         "AL-V(R)"          
#>  [91] "CA(L)"             "AL-D(L)"           "FBl1"             
#>  [94] "bL(R)"             "AL-VM7d(L)"        "AL-VA2(R)"        
#>  [97] "AL-VP1l(R)"        "NO3(R)"            "g2(R)"            
#> [100] "LX(R)"             "AL-VM7v(L)"        "PB(L7)"           
#> [103] "AL-VA1d(R)"        "NO"                "AL-VC1(R)"        
#> [106] "LAL(L)"            "IB112"             "AL(R)"            
#> [109] "AL-VC3m(R)"        "POC"               "EBr5"             
#> [112] "VES(L)"            "AL-VL2p(R)"        "vACA(R)"          
#> [115] "AL-DL2v(R)"        "AB(R)"             "NO(L)"            
#> [118] "AL-VL1(R)"         "CRE(-RUB)(L)"      "g5(R)"            
#> [121] "AL-DM2(R)"         "SIP(L)"            "AL-DL3(R)"        
#> [124] "hemibrain"         "b2(R)"             "MB(R)"            
#> [127] "EBr3am"            "SMP(L)"            "FBl6"             
#> [130] "AL-VM5v(R)"        "AOT(R)"            "RUB(R)"           
#> [133] "SPS(R)"            "VMNP"              "AL-VM4(R)"        
#> [136] "aL(R)"             "PB(L9)"            "CA(R)"            
#> [139] "AL-DM1(L)"         "AL-DL4(R)"         "SMP(R)"           
#> [142] "CRE(-ROB,-RUB)(R)" "g1(R)"             "AL-VP1m(R)"       
#> [145] "a1(R)"             "a'L(R)"            "AL-DC4(L)"        
#> [148] "AL-VP1d(R)"        "VLNP(R)"           "PB(R6)"           
#> [151] "GOR(R)"            "AL-DM4(R)"         "AB(L)"            
#> [154] "MB(L)"             "CRE(L)"            "NO2(R)"           
#> [157] "AL-DA4l(R)"        "AL-DM3(R)"         "AL-DP1m(R)"       
#> [160] "AL-VC5(R)"         "AMMC"              "EB"               
#> [163] "AL-DP1l(R)"        "AL-VM2(R)"         "b'2(R)"           
#> [166] "SAD(-AMMC)"        "ICL(R)"            "AL-DM1(R)"        
#> [169] "PB(R5)"            "NO1(L)"            "FBl8"             
#> [172] "b1(R)"             "BU(L)"             "SNP(R)"           
#> [175] "EBr3pw"            "a2(R)"             "GNG"              
#> [178] "a'3(R)"            "FB"                "FLA(R)"           
#> [181] "a'1(R)"            "AL-DA3(R)"         "AL-DC4(R)"        
#> [184] "AL-DL1(R)"         "PB(L5)"            "AL-DC1(L)"        
#> [187] "AL-DL5(L)"         "AL-DP1m(L)"        "ATL(L)"           
#> [190] "AL-DL4(L)"         "LO(R)"             "PB(L4)"           
#> [193] "lACA(R)"           "BU(R)"             "b'L(L)"           
#> [196] "AL-D(R)"           "AL-VP2(R)"         "CX"               
#> [199] "EBr6"              "PB(L6)"            "CAN(R)"           
#> [202] "dACA(R)"           "MB(+ACA)(R)"       "GF(R)"            
#> [205] "FBl3"              "PB(R2)"            "AL-DL5(R)"        
#> [208] "AL-VM7d(R)"        "mALT(R)"           "WED(R)"           
#> [211] "EBr3d"             "AL(L)"             "gL(R)"            
#> [214] "AL-DM5(R)"         "AL-VA4(R)"         "AL-VM3(R)"        
#> [217] "ME(R)"             "FBl7"              "AL-DM2(L)"        
#> [220] "FBl2"              "AL-DA4m(R)"        "b'L(R)"           
#> [223] "PB(R1)"            "VES(R)"            "PB(L3)"           
#> [226] "LAL(-GA)(R)"       "PB(L1)"            "AL-DA4m(L)"       
#> [229] "NO3(L)"            "AL-DC1(R)"         "b'1(R)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "SAD"               "LAL(-GA)(R)"       "LX(R)"            
#>   [4] "PB(R3)"            "GA(R)"             "a'L(L)"           
#>   [7] "b'2(R)"            "PB(L6)"            "a'2(R)"           
#>  [10] "SNP(R)"            "b1(R)"             "b'L(R)"           
#>  [13] "PLP(R)"            "ME(R)"             "ROB(R)"           
#>  [16] "PB(R2)"            "aL(L)"             "SMP(R)"           
#>  [19] "PRW"               "CA(R)"             "CA(L)"            
#>  [22] "SCL(R)"            "g4(R)"             "FBl3"             
#>  [25] "FBl9"              "VES(L)"            "SMP(L)"           
#>  [28] "PB(R9)"            "PB(L9)"            "LO(R)"            
#>  [31] "a1(R)"             "BU(L)"             "IB"               
#>  [34] "a'3(R)"            "vACA(R)"           "PB(R5)"           
#>  [37] "WED(R)"            "EPA(R)"            "PB(R1)"           
#>  [40] "PVLP(R)"           "LX(L)"             "SPS(R)"           
#>  [43] "FBl2"              "NO3(R)"            "INP"              
#>  [46] "CRE(R)"            "LH(R)"             "CAN(R)"           
#>  [49] "EPA(L)"            "GOR(R)"            "NO1(L)"           
#>  [52] "GOR(L)"            "EBr1"              "NO2(L)"           
#>  [55] "hemibrain"         "g2(R)"             "FLA(R)"           
#>  [58] "NO"                "NO(R)"             "MB(L)"            
#>  [61] "GNG"               "AOTU(R)"           "PB(R4)"           
#>  [64] "LAL(R)"            "SLP(R)"            "LOP(R)"           
#>  [67] "SIP(L)"            "ICL(L)"            "a3(R)"            
#>  [70] "PB(L2)"            "PED(R)"            "EBr5"             
#>  [73] "AB(R)"             "RUB(L)"            "AL(R)"            
#>  [76] "FB"                "PB"                "ATL(R)"           
#>  [79] "PB(L5)"            "CRE(-ROB,-RUB)(R)" "GC"               
#>  [82] "lACA(R)"           "PB(R7)"            "SPS(L)"           
#>  [85] "a2(R)"             "EBr6"              "dACA(R)"          
#>  [88] "b'L(L)"            "AOT(R)"            "RUB(R)"           
#>  [91] "PB(L3)"            "a'L(R)"            "BU(R)"            
#>  [94] "SNP(L)"            "b'1(R)"            "NO1(R)"           
#>  [97] "g3(R)"             "aL(R)"             "SAD(-AMMC)"       
#> [100] "VES(R)"            "MB(+ACA)(R)"       "AVLP(R)"          
#> [103] "EBr3am"            "LAL(L)"            "gL(L)"            
#> [106] "AME(R)"            "mALT(L)"           "GF(R)"            
#> [109] "FBl7"              "FBl8"              "PB(L4)"           
#> [112] "ATL(L)"            "EBr2r4"            "PB(L8)"           
#> [115] "EB"                "a'1(R)"            "g5(R)"            
#> [118] "FBl6"              "NO(L)"             "POC"              
#> [121] "g1(R)"             "CX"                "AL(L)"            
#> [124] "gL(R)"             "CRE(L)"            "FBl4"             
#> [127] "IPS(R)"            "bL(R)"             "VLNP(R)"          
#> [130] "mALT(R)"           "EBr3d"             "SIP(R)"           
#> [133] "EBr3pw"            "PB(L7)"            "FBl1"             
#> [136] "FBl5"              "b2(R)"             "PB(L1)"           
#> [139] "ICL(R)"            "NO2(R)"            "OL(R)"            
#> [142] "MB(R)"             "AB(L)"             "SCL(L)"           
#> [145] "PB(R6)"            "VMNP"              "bL(L)"            
#> [148] "AMMC"              "PB(R8)"            "NO3(L)"           
#> [151] "PENP"             
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "FBl8"              "MB(R)"             "ME(R)"            
#>   [4] "g5(R)"             "b'1(R)"            "AL-DM2(L)"        
#>   [7] "FBl3"              "PB(R4)"            "AL-VL2p(R)"       
#>  [10] "PED(R)"            "PLP(R)"            "AL-VC1(R)"        
#>  [13] "AB(L)"             "b'2(R)"            "LAL(L)"           
#>  [16] "EPA(L)"            "GNG"               "CAN(R)"           
#>  [19] "AL-VP5(R)"         "PB(R6)"            "ICL(L)"           
#>  [22] "ICL(R)"            "LOP(R)"            "EBr6"             
#>  [25] "a'L(R)"            "a'L(L)"            "VES(L)"           
#>  [28] "AL-VA4(R)"         "AL-DM5(L)"         "AL-DP1m(L)"       
#>  [31] "ROB(R)"            "CRE(-RUB)(L)"      "SNP(L)"           
#>  [34] "GA(R)"             "LAL(-GA)(R)"       "g4(R)"            
#>  [37] "b2(R)"             "SLP(R)"            "AL-VP1d(R)"       
#>  [40] "PB(L8)"            "SIP(L)"            "FBl1"             
#>  [43] "INP"               "PB(L3)"            "a'2(R)"           
#>  [46] "AL-VL1(R)"         "PB(R7)"            "NO3(L)"           
#>  [49] "AL-DL4(R)"         "AL-DP1l(R)"        "PB(R5)"           
#>  [52] "SPS(L)"            "gL(L)"             "PB(R3)"           
#>  [55] "GOR(R)"            "AL-VC5(R)"         "VLNP(R)"          
#>  [58] "AOT(R)"            "GOR(L)"            "aL(R)"            
#>  [61] "PB(L2)"            "a2(R)"             "PB(L7)"           
#>  [64] "LO(R)"             "EBr3am"            "SMP(L)"           
#>  [67] "NO1(R)"            "LAL(R)"            "FB"               
#>  [70] "hemibrain"         "PB(L1)"            "AL-DC1(L)"        
#>  [73] "SIP(R)"            "NO(R)"             "SAD(-AMMC)"       
#>  [76] "AL-DA3(R)"         "MB(L)"             "AL(L)"            
#>  [79] "FBl6"              "NO2(R)"            "AL-DL2v(R)"       
#>  [82] "AL-DM6(R)"         "WED(R)"            "AL-DM4(L)"        
#>  [85] "FBl2"              "CX"                "AL-VA2(R)"        
#>  [88] "AL-VA5(R)"         "AL-VM2(R)"         "AL-VP3(R)"        
#>  [91] "RUB(R)"            "a'3(R)"            "EPA(R)"           
#>  [94] "AL-D(L)"           "b'L(L)"            "AL(R)"            
#>  [97] "a'1(R)"            "b1(R)"             "aL(L)"            
#> [100] "AL-VA1v(R)"        "AL-VA7m(R)"        "AL-VP1l(R)"       
#> [103] "RUB(L)"            "AL-DA4l(R)"        "AL-DM4(R)"        
#> [106] "VMNP"              "PVLP(R)"           "NO(L)"            
#> [109] "AL-VC2(R)"         "g3(R)"             "AL-VA7l(R)"       
#> [112] "vACA(R)"           "mALT(R)"           "g1(R)"            
#> [115] "FBl4"              "CA(R)"             "BU(L)"            
#> [118] "EBr3d"             "AL-DM6(L)"         "ATL(R)"           
#> [121] "PB(R9)"            "FBl9"              "AL-DA2(R)"        
#> [124] "AL-VP1m(R)"        "PB(L4)"            "AL-DC4(R)"        
#> [127] "CRE(-ROB,-RUB)(R)" "a1(R)"             "AL-VM7v(L)"       
#> [130] "FB-column3"        "AME(R)"            "LX(L)"            
#> [133] "AL-DA4m(R)"        "mALT(L)"           "gL(R)"            
#> [136] "NO2(L)"            "EBr2r4"            "AL-DM1(L)"        
#> [139] "AL-DP1m(R)"        "ATL(L)"            "g2(R)"            
#> [142] "PENP"              "FLA(R)"            "AL-DC4(L)"        
#> [145] "AL-DM1(R)"         "AOTU(R)"           "SAD"              
#> [148] "SNP(R)"            "AL-DA3(L)"         "FBl7"             
#> [151] "PB(R2)"            "AL-VA6(R)"         "AL-DL5(L)"        
#> [154] "AL-DM3(L)"         "AL-DL5(R)"         "PB(L6)"           
#> [157] "AL-DM2(R)"         "AL-VM5d(R)"        "CRE(R)"           
#> [160] "AL-DA4m(L)"        "OL(R)"             "AL-VP2(R)"        
#> [163] "IB"                "SPS(R)"            "b'L(R)"           
#> [166] "AL-VM7d(L)"        "PRW"               "LX(R)"            
#> [169] "EB"                "AL-DC3(R)"         "GF(R)"            
#> [172] "EBr1"              "AVLP(R)"           "AL-VM5v(R)"       
#> [175] "AL-V(R)"           "lACA(R)"           "EBr3pw"           
#> [178] "AL-VP4(R)"         "AMMC"              "GC"               
#> [181] "SCL(R)"            "CRE(L)"            "LH(R)"            
#> [184] "AL-DM5(R)"         "AL-VC3l(R)"        "POC"              
#> [187] "PB"                "AL-DL4(L)"         "AL-VA3(R)"        
#> [190] "NO1(L)"            "IPS(R)"            "AL-DA1(R)"        
#> [193] "SCL(L)"            "PB(R1)"            "PB(L9)"           
#> [196] "MB(+ACA)(R)"       "AL-D(R)"           "AL-VA6(L)"        
#> [199] "AB(R)"             "AL-DM3(R)"         "SMP(R)"           
#> [202] "AL-DA2(L)"         "bL(L)"             "bL(R)"            
#> [205] "AL-VM4(R)"         "AL-VM7d(R)"        "FBl5"             
#> [208] "AL-DC1(R)"         "AL-DC2(R)"         "AL-DL2d(R)"       
#> [211] "AL-VC3m(R)"        "AL-VM1(R)"         "CA(L)"            
#> [214] "EBr5"              "dACA(R)"           "a3(R)"            
#> [217] "AL-DL3(R)"         "AL-VL2a(R)"        "NO3(R)"           
#> [220] "BU(R)"             "VES(R)"            "AL-DL1(R)"        
#> [223] "PB(R8)"            "AL-DC2(L)"         "AL-VM3(R)"        
#> [226] "AL-VC4(R)"         "PB(L5)"            "AL-VM7v(R)"       
#> [229] "NO"                "AL-VA1d(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-DM3(L)"         "AL-DM4(L)"         "AL-DL5(R)"        
#>   [4] "LOP(R)"            "SPS(L)"            "MB(+ACA)(R)"      
#>   [7] "AL-DA4l(R)"        "ATL(R)"            "a'3(R)"           
#>  [10] "AL-VA7m(R)"        "AL-VP1l(R)"        "AOTU(R)"          
#>  [13] "SIP(L)"            "MB(L)"             "EBr1"             
#>  [16] "AL-DL1(R)"         "BU(L)"             "PB(R9)"           
#>  [19] "SAD(-AMMC)"        "FB"                "AL-DM2(L)"        
#>  [22] "NO3(L)"            "AL-VM1(R)"         "mALT(L)"          
#>  [25] "SNP(L)"            "AL-VA6(L)"         "FBl8"             
#>  [28] "AL-DA2(R)"         "AL-DL4(R)"         "g3(R)"            
#>  [31] "b2(R)"             "g4(R)"             "VLNP(R)"          
#>  [34] "CX"                "CAN(R)"            "SMP(R)"           
#>  [37] "b'1(R)"            "mALT(R)"           "EBr3d"            
#>  [40] "SIP(R)"            "PB(R1)"            "FBl9"             
#>  [43] "CRE(-ROB,-RUB)(R)" "NO2(L)"            "hemibrain"        
#>  [46] "AL-VA3(R)"         "EBr3am"            "VES(L)"           
#>  [49] "MB(R)"             "AL-VP5(R)"         "PB(R5)"           
#>  [52] "SCL(R)"            "EBr6"              "RUB(L)"           
#>  [55] "aL(R)"             "NO"                "LAL(-GA)(R)"      
#>  [58] "a'L(R)"            "FBl5"              "AL-VP2(R)"        
#>  [61] "ATL(L)"            "gL(R)"             "AL-DA1(R)"        
#>  [64] "AL-VA1d(R)"        "AL-VC1(R)"         "AB(L)"            
#>  [67] "AL-DA2(L)"         "a'1(R)"            "AL-VM7d(L)"       
#>  [70] "FBl1"              "g2(R)"             "FB-column3"       
#>  [73] "AL-VM7v(R)"        "PB(R6)"            "PED(R)"           
#>  [76] "GC"                "VMNP"              "AL-VM3(R)"        
#>  [79] "PB(L8)"            "PB(L7)"            "AL-VA1v(R)"       
#>  [82] "SPS(R)"            "EBr5"              "AL-DA3(L)"        
#>  [85] "SLP(R)"            "AL-VM7v(L)"        "ROB(R)"           
#>  [88] "NO(R)"             "AL-VP3(R)"         "vACA(R)"          
#>  [91] "AL-DP1m(L)"        "AL-DM5(R)"         "CA(L)"            
#>  [94] "GOR(L)"            "CRE(R)"            "AL-DC1(L)"        
#>  [97] "AL-DM1(L)"         "AL-DM1(R)"         "WED(R)"           
#> [100] "AL-VC3m(R)"        "PLP(R)"            "AL-D(L)"          
#> [103] "PVLP(R)"           "FBl3"              "PB(L2)"           
#> [106] "PRW"               "FLA(R)"            "AL-V(R)"          
#> [109] "lACA(R)"           "FBl2"              "AL-DC4(R)"        
#> [112] "PB(R7)"            "AL-VA2(R)"         "AL-VL1(R)"        
#> [115] "AL-VL2p(R)"        "RUB(R)"            "gL(L)"            
#> [118] "AL-DL2d(R)"        "AL-VC3l(R)"        "AL-VM4(R)"        
#> [121] "ICL(L)"            "AL-VL2a(R)"        "AL-VM7d(R)"       
#> [124] "AL-VA7l(R)"        "PENP"              "b'2(R)"           
#> [127] "NO1(L)"            "PB(L1)"            "OL(R)"            
#> [130] "LO(R)"             "LX(L)"             "a'2(R)"           
#> [133] "SNP(R)"            "FBl7"              "AME(R)"           
#> [136] "bL(R)"             "LAL(L)"            "AL-DM5(L)"        
#> [139] "ME(R)"             "FBl4"              "AL-DA3(R)"        
#> [142] "AL-DL3(R)"         "POC"               "PB(L3)"           
#> [145] "IPS(R)"            "AL-VC4(R)"         "AL-VP1d(R)"       
#> [148] "SCL(L)"            "LX(R)"             "CA(R)"            
#> [151] "b'L(L)"            "a'L(L)"            "AL-DM4(R)"        
#> [154] "AL-DC1(R)"         "GA(R)"             "AL-DA4m(L)"       
#> [157] "AVLP(R)"           "NO2(R)"            "NO(L)"            
#> [160] "AL-DL5(L)"         "AL-VA6(R)"         "SAD"              
#> [163] "NO1(R)"            "AL-VC2(R)"         "LH(R)"            
#> [166] "GOR(R)"            "AL(R)"             "AL-DM2(R)"        
#> [169] "PB(L9)"            "b1(R)"             "AL-VP1m(R)"       
#> [172] "EPA(R)"            "AL(L)"             "PB(L6)"           
#> [175] "AL-DC3(R)"         "AL-DM3(R)"         "AL-DM6(L)"        
#> [178] "CRE(L)"            "GNG"               "AL-VM2(R)"        
#> [181] "SMP(L)"            "PB(R8)"            "PB(R2)"           
#> [184] "PB(R3)"            "b'L(R)"            "EPA(L)"           
#> [187] "AL-DP1l(R)"        "AL-VA4(R)"         "PB(L5)"           
#> [190] "AL-D(R)"           "AL-VM5d(R)"        "BU(R)"            
#> [193] "AL-DL4(L)"         "aL(L)"             "IB"               
#> [196] "a1(R)"             "EBr3pw"            "AL-VC5(R)"        
#> [199] "AL-VP4(R)"         "AOT(R)"            "a2(R)"            
#> [202] "NO3(R)"            "AB(R)"             "AL-DL2v(R)"       
#> [205] "AL-DC4(L)"         "FBl6"              "AL-DA4m(R)"       
#> [208] "bL(L)"             "LAL(R)"            "AL-VM5v(R)"       
#> [211] "dACA(R)"           "AL-DM6(R)"         "AL-DP1m(R)"       
#> [214] "PB"                "AL-DC2(L)"         "CRE(-RUB)(L)"     
#> [217] "AMMC"              "g5(R)"             "VES(R)"           
#> [220] "PB(R4)"            "a3(R)"             "PB(L4)"           
#> [223] "AL-DC2(R)"         "AL-VA5(R)"         "INP"              
#> [226] "GF(R)"             "g1(R)"             "EBr2r4"           
#> [229] "EB"                "ICL(R)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                               
#> [1,] "CONNECT" "/api"                             
#> [2,] "POST"    "/api/npexplorer/simpleconnections"
#> [3,] "POST"    "/api/v:ver/raw/cypher/transaction"
#> [4,] "GET"     "/api/v:ver/dbmeta/datasets"       
#> [5,] "POST"    "/api/custom/custom"               
#> [6,] "GET"     "/profile"                         
#>      name                                                                                             
#> [1,] "github.com/labstack/echo.glob..func1"                                                           
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"                  
#> [6,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                         

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }