Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $fib19
#> $fib19$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $fib19$uuid
#> [1] "6c1b926af14f43cbbcf85da97c46fb2b"
#> 
#> $fib19$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $fib19$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $fib19$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $fib19$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "MB(+ACA)(R)"       "PB(R2)"            "EPA(L)"           
#>   [4] "AL-VL1(R)"         "a2(R)"             "NO3(R)"           
#>   [7] "b'L(L)"            "AL-DL5(R)"         "AL-VC1(R)"        
#>  [10] "GC"                "AL-VM1(R)"         "AL-VM7v(R)"       
#>  [13] "NO2(L)"            "EBr6"              "PB(L6)"           
#>  [16] "bL(R)"             "LX(R)"             "LAL(L)"           
#>  [19] "a'1(R)"            "AL-DA4l(R)"        "AL-VA5(R)"        
#>  [22] "EBr3am"            "AMMC"              "aL(R)"            
#>  [25] "gL(L)"             "AB(L)"             "b'1(R)"           
#>  [28] "AL-DC1(L)"         "AL(R)"             "AL-VA1d(R)"       
#>  [31] "AL-VA1v(R)"        "VMNP"              "IB112"            
#>  [34] "AL-VC3l(R)"        "AL-VP2(R)"         "PB(L9)"           
#>  [37] "AB(R)"             "NO(L)"             "PB(L2)"           
#>  [40] "PB(R4)"            "AL-VP1d(R)"        "SMP(R)"           
#>  [43] "gL(R)"             "CAN(R)"            "FB-column3"       
#>  [46] "AL-VC2(R)"         "AL-VM5d(R)"        "b'2(R)"           
#>  [49] "aL(L)"             "PB(L4)"            "AL-DC3(R)"        
#>  [52] "AL-VM7d(R)"        "EB"                "SIP(R)"           
#>  [55] "AL-DM6(R)"         "a1(R)"             "AL-DM4(R)"        
#>  [58] "b'L(R)"            "PLP(R)"            "ATL(R)"           
#>  [61] "FLA(R)"            "PVLP(R)"           "hemibrain"        
#>  [64] "EBr3d"             "AL-DA2(L)"         "EBr3pw"           
#>  [67] "INP"               "AL-VC3m(R)"        "AL-DM2(R)"        
#>  [70] "AL-VA3(R)"         "g3(R)"             "SIP(L)"           
#>  [73] "NO(R)"             "CRE(-RUB)(L)"      "AL-VP4(R)"        
#>  [76] "AOT(R)"            "PB(R7)"            "SPS(L)"           
#>  [79] "AL-VA6(L)"         "NO3(L)"            "GNG"              
#>  [82] "AL-VA6(R)"         "SMP(L)"            "AL-DA3(L)"        
#>  [85] "SAD(-AMMC)"        "ICL(R)"            "BU(L)"            
#>  [88] "SLP(R)"            "AL-V(R)"           "AL-DM6(L)"        
#>  [91] "LH(R)"             "PB(R3)"            "AL-D(L)"          
#>  [94] "FBl7"              "AL-VP1l(R)"        "PB(R9)"           
#>  [97] "bL(L)"             "AL-DP1l(R)"        "PB(L8)"           
#> [100] "g4(R)"             "LO(R)"             "FBl4"             
#> [103] "AL-DA2(R)"         "AL-DC4(R)"         "AL(L)"            
#> [106] "AL-DC2(L)"         "AL-DM5(L)"         "mALT(L)"          
#> [109] "b2(R)"             "FBl5"              "b1(R)"            
#> [112] "AL-DL4(R)"         "RUB(R)"            "SNP(R)"           
#> [115] "AL-DA1(R)"         "AL-DL2d(R)"        "g1(R)"            
#> [118] "AL-DM1(L)"         "VES(L)"            "AL-VM3(R)"        
#> [121] "PB(L5)"            "LAL(-GA)(R)"       "MB(L)"            
#> [124] "CRE(L)"            "RUB(L)"            "AL-VM5v(R)"       
#> [127] "AL-DM2(L)"         "PB(L7)"            "FBl2"             
#> [130] "AL-DM1(R)"         "AL-DM5(R)"         "WED(R)"           
#> [133] "AL-DA4m(L)"        "AL-DP1m(L)"        "ROB(R)"           
#> [136] "VES(R)"            "GOR(R)"            "NO2(R)"           
#> [139] "AL-VL2a(R)"        "IB"                "a'3(R)"           
#> [142] "AOTU(R)"           "PRW"               "EBr5"             
#> [145] "LX(L)"             "AL-D(R)"           "PB(R5)"           
#> [148] "g2(R)"             "PB(L1)"            "EPA(R)"           
#> [151] "AL-VM2(R)"         "AL-VP3(R)"         "GA(R)"            
#> [154] "vACA(R)"           "PB"                "AL-DC4(L)"        
#> [157] "SNP(L)"            "FBl6"              "LAL(R)"           
#> [160] "AL-VA2(R)"         "AL-DA3(R)"         "AL-DA4m(R)"       
#> [163] "AL-VC4(R)"         "POC"               "SCL(L)"           
#> [166] "ATL(L)"            "AL-DL3(R)"         "FBl3"             
#> [169] "CRE(R)"            "NO1(R)"            "IPS(R)"           
#> [172] "VLNP(R)"           "SAD"               "PED(R)"           
#> [175] "EBr1"              "a'L(R)"            "AL-DC2(R)"        
#> [178] "AL-DP1m(R)"        "g5(R)"             "FBl9"             
#> [181] "CA(L)"             "a'2(R)"            "AL-VM7v(L)"       
#> [184] "AL-DC1(R)"         "AL-VP5(R)"         "MB(R)"            
#> [187] "dACA(R)"           "AL-DL4(L)"         "AL-DM3(L)"        
#> [190] "AVLP(R)"           "AL-VL2p(R)"        "a3(R)"            
#> [193] "AL-DL1(R)"         "CRE(-ROB,-RUB)(R)" "EBr2r4"           
#> [196] "AL-VC5(R)"         "CX"                "CA(R)"            
#> [199] "GOR(L)"            "PB(R8)"            "AL-VM7d(L)"       
#> [202] "NO"                "PB(R1)"            "FBl8"             
#> [205] "SCL(R)"            "a'L(L)"            "NO1(L)"           
#> [208] "ME(R)"             "OL(R)"             "AL-VM4(R)"        
#> [211] "AL-VP1m(R)"        "AL-VA7l(R)"        "GF(R)"            
#> [214] "lACA(R)"           "AL-DM4(L)"         "AL-DM3(R)"        
#> [217] "FB"                "FBl1"              "AME(R)"           
#> [220] "AL-VA7m(R)"        "PENP"              "PB(R6)"           
#> [223] "mALT(R)"           "BU(R)"             "ICL(L)"           
#> [226] "AL-VA4(R)"         "PB(L3)"            "SPS(R)"           
#> [229] "AL-DL5(L)"         "AL-DL2v(R)"        "LOP(R)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "NO3(L)"            "CA(L)"             "b'L(L)"           
#>   [4] "AB(L)"             "mALT(L)"           "IB"               
#>   [7] "g1(R)"             "PB(R9)"            "gL(R)"            
#>  [10] "MB(+ACA)(R)"       "PB(R2)"            "EBr5"             
#>  [13] "ICL(R)"            "LX(L)"             "b'1(R)"           
#>  [16] "ATL(R)"            "NO2(L)"            "PED(R)"           
#>  [19] "dACA(R)"           "LX(R)"             "PVLP(R)"          
#>  [22] "FBl9"              "PENP"              "a'2(R)"           
#>  [25] "PB(R1)"            "SPS(L)"            "EPA(R)"           
#>  [28] "NO(R)"             "SNP(L)"            "FBl1"             
#>  [31] "AB(R)"             "lACA(R)"           "ROB(R)"           
#>  [34] "PB(L6)"            "EBr3pw"            "FBl5"             
#>  [37] "LAL(R)"            "VES(L)"            "GA(R)"            
#>  [40] "GOR(L)"            "PLP(R)"            "BU(R)"            
#>  [43] "FBl3"              "MB(L)"             "ICL(L)"           
#>  [46] "EPA(L)"            "WED(R)"            "EBr6"             
#>  [49] "FBl2"              "CA(R)"             "FBl4"             
#>  [52] "IPS(R)"            "MB(R)"             "SCL(L)"           
#>  [55] "PB(L5)"            "SAD"               "NO1(L)"           
#>  [58] "PB(L7)"            "SLP(R)"            "CRE(-ROB,-RUB)(R)"
#>  [61] "FLA(R)"            "FBl7"              "b1(R)"            
#>  [64] "GNG"               "NO(L)"             "VLNP(R)"          
#>  [67] "PB(R7)"            "PB"                "AL(L)"            
#>  [70] "SAD(-AMMC)"        "LO(R)"             "EBr2r4"           
#>  [73] "vACA(R)"           "b'2(R)"            "ME(R)"            
#>  [76] "gL(L)"             "g3(R)"             "b2(R)"            
#>  [79] "g4(R)"             "AVLP(R)"           "AME(R)"           
#>  [82] "a'1(R)"            "PB(L4)"            "a2(R)"            
#>  [85] "CRE(R)"            "g2(R)"             "PB(R8)"           
#>  [88] "EB"                "a'3(R)"            "SCL(R)"           
#>  [91] "PB(L9)"            "bL(L)"             "INP"              
#>  [94] "PB(R5)"            "BU(L)"             "PB(L8)"           
#>  [97] "VMNP"              "PB(R4)"            "RUB(R)"           
#> [100] "SPS(R)"            "CAN(R)"            "CX"               
#> [103] "PRW"               "EBr1"              "PB(L1)"           
#> [106] "SNP(R)"            "FBl8"              "NO3(R)"           
#> [109] "PB(R3)"            "RUB(L)"            "AOT(R)"           
#> [112] "PB(R6)"            "CRE(L)"            "a3(R)"            
#> [115] "EBr3am"            "a1(R)"             "g5(R)"            
#> [118] "OL(R)"             "AL(R)"             "SMP(L)"           
#> [121] "SIP(L)"            "LAL(-GA)(R)"       "mALT(R)"          
#> [124] "EBr3d"             "POC"               "PB(L2)"           
#> [127] "NO2(R)"            "LOP(R)"            "GF(R)"            
#> [130] "NO1(R)"            "SIP(R)"            "VES(R)"           
#> [133] "LH(R)"             "FBl6"              "ATL(L)"           
#> [136] "AMMC"              "PB(L3)"            "AOTU(R)"          
#> [139] "hemibrain"         "b'L(R)"            "NO"               
#> [142] "GC"                "FB"                "aL(L)"            
#> [145] "GOR(R)"            "SMP(R)"            "LAL(L)"           
#> [148] "aL(R)"             "a'L(R)"            "bL(R)"            
#> [151] "a'L(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DM5(L)"         "AME(R)"            "AL-DL4(R)"        
#>   [4] "AL-VA5(R)"         "CRE(-ROB,-RUB)(R)" "PB(R5)"           
#>   [7] "SMP(R)"            "EBr2r4"            "AL-VA6(L)"        
#>  [10] "SNP(L)"            "NO(L)"             "AL-VC4(R)"        
#>  [13] "AL-VP1m(R)"        "SCL(R)"            "a'2(R)"           
#>  [16] "hemibrain"         "FBl8"              "AVLP(R)"          
#>  [19] "AL-VP3(R)"         "a'1(R)"            "AL-DL2v(R)"       
#>  [22] "AL-DM5(R)"         "EBr3am"            "NO2(L)"           
#>  [25] "EPA(R)"            "FBl3"              "EPA(L)"           
#>  [28] "AL-DL4(L)"         "NO3(L)"            "AL-VM7d(R)"       
#>  [31] "PED(R)"            "GC"                "AL-VP1d(R)"       
#>  [34] "g1(R)"             "aL(R)"             "PB"               
#>  [37] "AL-VA1v(R)"        "FBl4"              "AL-DM2(R)"        
#>  [40] "AL-VA4(R)"         "PB(L1)"            "SIP(R)"           
#>  [43] "AL-DM1(R)"         "AL-DM2(L)"         "AL-DC2(R)"        
#>  [46] "GA(R)"             "LX(R)"             "AL-VA3(R)"        
#>  [49] "AL-VL2p(R)"        "FB"                "AL-DM6(R)"        
#>  [52] "a'L(R)"            "AL-DC1(L)"         "NO(R)"            
#>  [55] "LAL(R)"            "AL-DM4(R)"         "AL-VC2(R)"        
#>  [58] "dACA(R)"           "g5(R)"             "a3(R)"            
#>  [61] "b'L(L)"            "PB(L4)"            "AL-VP4(R)"        
#>  [64] "WED(R)"            "b2(R)"             "LH(R)"            
#>  [67] "b1(R)"             "LX(L)"             "AL-VC5(R)"        
#>  [70] "AL-VL1(R)"         "AL-DC3(R)"         "gL(L)"            
#>  [73] "AL(R)"             "AL-VA1d(R)"        "EBr5"             
#>  [76] "LO(R)"             "ATL(R)"            "AL-DL3(R)"        
#>  [79] "SPS(R)"            "EBr3pw"            "PB(L7)"           
#>  [82] "SMP(L)"            "AB(L)"             "AL-VM7v(R)"       
#>  [85] "NO1(R)"            "AL-DC2(L)"         "AL-DL5(L)"        
#>  [88] "ROB(R)"            "CRE(L)"            "AL-VM1(R)"        
#>  [91] "AL-DA4m(R)"        "AL-VM5d(R)"        "AL-DM6(L)"        
#>  [94] "AL-VM7d(L)"        "gL(R)"             "CA(R)"            
#>  [97] "AL-DL1(R)"         "AOT(R)"            "PENP"             
#> [100] "CA(L)"             "PVLP(R)"           "FBl2"             
#> [103] "AL-DC1(R)"         "PLP(R)"            "MB(+ACA)(R)"      
#> [106] "FBl6"              "VLNP(R)"           "BU(L)"            
#> [109] "AL-D(L)"           "ICL(L)"            "AL-DM3(R)"        
#> [112] "AL-VM5v(R)"        "AL-DC4(L)"         "ATL(L)"           
#> [115] "AL-DP1m(R)"        "AL-VM2(R)"         "b'1(R)"           
#> [118] "PB(R3)"            "GOR(L)"            "CRE(-RUB)(L)"     
#> [121] "FB-column3"        "AL-VM7v(L)"        "FBl5"             
#> [124] "GOR(R)"            "AL-DL5(R)"         "PB(R9)"           
#> [127] "AL-DA1(R)"         "IB"                "SCL(L)"           
#> [130] "vACA(R)"           "GF(R)"             "CRE(R)"           
#> [133] "AL-DA3(L)"         "VES(L)"            "AL-DA2(R)"        
#> [136] "AMMC"              "GNG"               "CAN(R)"           
#> [139] "AL-DC4(R)"         "AL-VC3m(R)"        "AL-VM4(R)"        
#> [142] "AL-VC3l(R)"        "SAD"               "EBr6"             
#> [145] "ME(R)"             "AL-DM3(L)"         "aL(L)"            
#> [148] "AL-D(R)"           "g2(R)"             "VMNP"             
#> [151] "AL-VP5(R)"         "AL-DP1m(L)"        "IPS(R)"           
#> [154] "AL-VP2(R)"         "CX"                "PB(L8)"           
#> [157] "PB(L5)"            "mALT(L)"           "SPS(L)"           
#> [160] "AL-VA6(R)"         "LOP(R)"            "bL(R)"            
#> [163] "a2(R)"             "lACA(R)"           "BU(R)"            
#> [166] "AL-DM4(L)"         "VES(R)"            "FBl1"             
#> [169] "AB(R)"             "a1(R)"             "PB(R8)"           
#> [172] "AL-DA3(R)"         "PB(L3)"            "PB(L9)"           
#> [175] "AL-DA4l(R)"        "AL-VP1l(R)"        "LAL(L)"           
#> [178] "SIP(L)"            "EB"                "EBr3d"            
#> [181] "g4(R)"             "AL-DL2d(R)"        "NO1(L)"           
#> [184] "AL(L)"             "AL-DA4m(L)"        "AL-DM1(L)"        
#> [187] "PB(L6)"            "a'L(L)"            "PB(R6)"           
#> [190] "PB(R7)"            "AL-DA2(L)"         "MB(R)"            
#> [193] "AL-VC1(R)"         "FBl7"              "PB(L2)"           
#> [196] "bL(L)"             "INP"               "POC"              
#> [199] "b'2(R)"            "NO3(R)"            "MB(L)"            
#> [202] "SAD(-AMMC)"        "FBl9"              "NO2(R)"           
#> [205] "PB(R1)"            "mALT(R)"           "SNP(R)"           
#> [208] "ICL(R)"            "AL-VA7m(R)"        "AL-VL2a(R)"       
#> [211] "RUB(R)"            "g3(R)"             "RUB(L)"           
#> [214] "AL-VA7l(R)"        "PRW"               "LAL(-GA)(R)"      
#> [217] "OL(R)"             "AL-DP1l(R)"        "AL-VA2(R)"        
#> [220] "NO"                "SLP(R)"            "a'3(R)"           
#> [223] "AOTU(R)"           "b'L(R)"            "PB(R4)"           
#> [226] "AL-V(R)"           "EBr1"              "FLA(R)"           
#> [229] "PB(R2)"            "AL-VM3(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "SCL(R)"            "AL-DC2(R)"         "AL-VA7l(R)"       
#>   [4] "PB(R7)"            "gL(L)"             "NO2(L)"           
#>   [7] "AL-DM6(L)"         "a3(R)"             "LAL(L)"           
#>  [10] "bL(L)"             "AL-VP5(R)"         "SMP(L)"           
#>  [13] "FBl1"              "FBl3"              "AL-VP1m(R)"       
#>  [16] "AL-VM3(R)"         "AL-DL5(L)"         "AL-DP1m(L)"       
#>  [19] "FBl6"              "AL-DL1(R)"         "PRW"              
#>  [22] "dACA(R)"           "AL-D(L)"           "AL-VA6(L)"        
#>  [25] "AL-DM1(R)"         "AL-DP1m(R)"        "AL-VM5d(R)"       
#>  [28] "SNP(R)"            "AL-VM7v(L)"        "AL-DL2v(R)"       
#>  [31] "AL-DM6(R)"         "CA(L)"             "AL-DM2(L)"        
#>  [34] "FBl9"              "AL-VL2a(R)"        "CRE(-ROB,-RUB)(R)"
#>  [37] "AL-VA4(R)"         "MB(L)"             "GOR(R)"           
#>  [40] "vACA(R)"           "PB(R9)"            "AL-VC2(R)"        
#>  [43] "NO3(R)"            "AL-V(R)"           "AB(L)"            
#>  [46] "AL-DM4(R)"         "SPS(R)"            "mALT(R)"          
#>  [49] "GC"                "gL(R)"             "GA(R)"            
#>  [52] "aL(R)"             "OL(R)"             "PB(R6)"           
#>  [55] "AL-DA3(R)"         "PB(L5)"            "PLP(R)"           
#>  [58] "ME(R)"             "AL-VC1(R)"         "SCL(L)"           
#>  [61] "AMMC"              "FB"                "FLA(R)"           
#>  [64] "ICL(L)"            "MB(+ACA)(R)"       "WED(R)"           
#>  [67] "PB(L9)"            "AVLP(R)"           "NO(L)"            
#>  [70] "LX(R)"             "AL-DM4(L)"         "EBr3pw"           
#>  [73] "CA(R)"             "g2(R)"             "EBr1"             
#>  [76] "NO2(R)"            "AL-VC3m(R)"        "b2(R)"            
#>  [79] "NO1(R)"            "PB(R8)"            "a'2(R)"           
#>  [82] "VMNP"              "a1(R)"             "EB"               
#>  [85] "aL(L)"             "VLNP(R)"           "AOT(R)"           
#>  [88] "b1(R)"             "AL-DA4l(R)"        "b'1(R)"           
#>  [91] "bL(R)"             "AL-VP3(R)"         "LOP(R)"           
#>  [94] "g3(R)"             "EBr5"              "VES(L)"           
#>  [97] "AL-VP4(R)"         "BU(L)"             "AL-DL4(L)"        
#> [100] "AL-DL2d(R)"        "CRE(R)"            "AL-DM5(L)"        
#> [103] "AL-VM7d(L)"        "AL-DM2(R)"         "NO(R)"            
#> [106] "AL-DL5(R)"         "AL-VA2(R)"         "AL-VM7v(R)"       
#> [109] "NO1(L)"            "SNP(L)"            "LAL(R)"           
#> [112] "AL-DP1l(R)"        "AL-VA1v(R)"        "IB"               
#> [115] "SIP(L)"            "SAD(-AMMC)"        "CAN(R)"           
#> [118] "AL-DA4m(R)"        "AL-VM7d(R)"        "NO3(L)"           
#> [121] "FBl4"              "AL-DL4(R)"         "AL-VC4(R)"        
#> [124] "PB"                "SIP(R)"            "CRE(-RUB)(L)"     
#> [127] "AL-VP1d(R)"        "AL-VA1d(R)"        "a'3(R)"           
#> [130] "PENP"              "GOR(L)"            "PB(R3)"           
#> [133] "AL-DM3(R)"         "AL-DA2(L)"         "mALT(L)"          
#> [136] "AL-VM2(R)"         "PB(L7)"            "LO(R)"            
#> [139] "a'L(L)"            "AL-DA1(R)"         "CRE(L)"           
#> [142] "AL-VA7m(R)"        "AL-VM1(R)"         "b'2(R)"           
#> [145] "AL-DA4m(L)"        "IPS(R)"            "AL-VC3l(R)"       
#> [148] "GF(R)"             "b'L(R)"            "FBl5"             
#> [151] "AL-DC3(R)"         "PB(L1)"            "g4(R)"            
#> [154] "PVLP(R)"           "CX"                "g1(R)"            
#> [157] "FB-column3"        "EBr3am"            "PB(L3)"           
#> [160] "AL-DM3(L)"         "PB(R2)"            "LX(L)"            
#> [163] "AL-DC4(L)"         "SLP(R)"            "AL-VC5(R)"        
#> [166] "ATL(R)"            "INP"               "a'L(R)"           
#> [169] "VES(R)"            "FBl8"              "LH(R)"            
#> [172] "PB(R1)"            "PB(L4)"            "AL-D(R)"          
#> [175] "AOTU(R)"           "EBr6"              "EBr3d"            
#> [178] "EPA(R)"            "MB(R)"             "AL-VP1l(R)"       
#> [181] "AL-DM1(L)"         "FBl2"              "ICL(R)"           
#> [184] "AL-DA2(R)"         "a2(R)"             "PB(L8)"           
#> [187] "AL(L)"             "AL-DC2(L)"         "AB(R)"            
#> [190] "AL-DM5(R)"         "PB(R4)"            "AL-DC1(R)"        
#> [193] "AL-DL3(R)"         "AL-VL1(R)"         "EBr2r4"           
#> [196] "BU(R)"             "a'1(R)"            "g5(R)"            
#> [199] "AL-VL2p(R)"        "SMP(R)"            "hemibrain"        
#> [202] "GNG"               "AL(R)"             "AL-VA3(R)"        
#> [205] "RUB(L)"            "b'L(L)"            "AL-VA6(R)"        
#> [208] "AME(R)"            "PB(R5)"            "ATL(L)"           
#> [211] "lACA(R)"           "LAL(-GA)(R)"       "AL-DC1(L)"        
#> [214] "PB(L6)"            "PED(R)"            "EPA(L)"           
#> [217] "NO"                "RUB(R)"            "PB(L2)"           
#> [220] "SAD"               "AL-DA3(L)"         "AL-DC4(R)"        
#> [223] "AL-VP2(R)"         "SPS(L)"            "ROB(R)"           
#> [226] "AL-VA5(R)"         "AL-VM4(R)"         "FBl7"             
#> [229] "POC"               "AL-VM5v(R)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method    path                                      
#> [1,] "GET"     "/api/dbmeta/datasets"                    
#> [2,] "POST"    "/api/v:ver/npexplorer/neuronmeta"        
#> [3,] "GET"     "/api/v:ver/npexplorer/commonconnectivity"
#> [4,] "POST"    "/api/v:ver/npexplorer/commonconnectivity"
#> [5,] "GET"     "/api/npexplorer/completeness"            
#> [6,] "OPTIONS" "/api"                                    
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatasets-fm"            
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCompleteness-fm"
#> [6,] "github.com/labstack/echo.glob..func1"                                                      
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }