Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "EBr3pw"            "IB112"             "VES(R)"           
#>   [4] "PB(R3)"            "GNG"               "AL-D(R)"          
#>   [7] "PB(R7)"            "AVLP(R)"           "OL(R)"            
#>  [10] "AL-VC3l(R)"        "a'L(R)"            "PB(R1)"           
#>  [13] "bL(R)"             "AL-DL2d(R)"        "NO1(R)"           
#>  [16] "AL-DC2(L)"         "AL-DC4(L)"         "CRE(-RUB)(L)"     
#>  [19] "AL-DP1m(L)"        "AL-DL5(L)"         "gL(R)"            
#>  [22] "AL-DM6(R)"         "LAL(-GA)(R)"       "aL(R)"            
#>  [25] "PB(R4)"            "AL-DA3(R)"         "AL-VA4(R)"        
#>  [28] "AL-VP3(R)"         "SAD(-AMMC)"        "IB"               
#>  [31] "b'L(R)"            "AL-DC1(L)"         "SIP(R)"           
#>  [34] "PB(L4)"            "AL-DM1(R)"         "ATL(R)"           
#>  [37] "AL-DA1(R)"         "AL-DA2(R)"         "vACA(R)"          
#>  [40] "b'2(R)"            "SPS(R)"            "a'1(R)"           
#>  [43] "ICL(L)"            "IPS(R)"            "AL(R)"            
#>  [46] "b2(R)"             "PB(L8)"            "AL-VA6(L)"        
#>  [49] "AL-DC2(R)"         "AL-DM5(R)"         "AL-VA5(R)"        
#>  [52] "AL-VM5v(R)"        "PED(R)"            "FLA(R)"           
#>  [55] "AL-DM1(L)"         "NO"                "SNP(L)"           
#>  [58] "AL-DC4(R)"         "AL-VP5(R)"         "EBr5"             
#>  [61] "LO(R)"             "CA(R)"             "AL-VP1m(R)"       
#>  [64] "ATL(L)"            "SIP(L)"            "EB"               
#>  [67] "PB(R2)"            "AL-DA2(L)"         "AL-DA4m(L)"       
#>  [70] "AME(R)"            "EBr6"              "PB"               
#>  [73] "AL-V(R)"           "POC"               "PLP(R)"           
#>  [76] "AL-DL4(R)"         "PENP"              "FBl6"             
#>  [79] "EPA(L)"            "NO(L)"             "a1(R)"            
#>  [82] "MB(L)"             "ME(R)"             "FBl9"             
#>  [85] "LH(R)"             "g1(R)"             "GOR(R)"           
#>  [88] "AL-VP4(R)"         "GC"                "CRE(R)"           
#>  [91] "AL-VL1(R)"         "AL-VP2(R)"         "EBr1"             
#>  [94] "gL(L)"             "AL-VC4(R)"         "AMMC"             
#>  [97] "AOTU(R)"           "b1(R)"             "EBr3d"            
#> [100] "LOP(R)"            "dACA(R)"           "PB(L7)"           
#> [103] "AL-DC1(R)"         "AL-VC2(R)"         "WED(R)"           
#> [106] "PB(R8)"            "AL-D(L)"           "CAN(R)"           
#> [109] "AL-DM2(R)"         "FB"                "g5(R)"            
#> [112] "MB(+ACA)(R)"       "PB(L1)"            "AL-VA3(R)"        
#> [115] "NO2(L)"            "a2(R)"             "AL(L)"            
#> [118] "PB(R9)"            "SAD"               "AL-DL4(L)"        
#> [121] "VES(L)"            "AL-VM7v(R)"        "AL-DM4(L)"        
#> [124] "NO(R)"             "AL-VA7l(R)"        "AL-VP1l(R)"       
#> [127] "GF(R)"             "PRW"               "LX(R)"            
#> [130] "BU(R)"             "FBl3"              "a'L(L)"           
#> [133] "AL-DL3(R)"         "AL-VM4(R)"         "EBr3am"           
#> [136] "PB(L6)"            "bL(L)"             "AB(R)"            
#> [139] "aL(L)"             "b'L(L)"            "SMP(R)"           
#> [142] "PB(L9)"            "AL-VL2p(R)"        "SMP(L)"           
#> [145] "RUB(L)"            "a3(R)"             "PB(R5)"           
#> [148] "AB(L)"             "EBr2r4"            "PVLP(R)"          
#> [151] "SLP(R)"            "MB(R)"             "AL-VA6(R)"        
#> [154] "VLNP(R)"           "AL-DL1(R)"         "RUB(R)"           
#> [157] "SPS(L)"            "AL-DM6(L)"         "AL-DA4l(R)"       
#> [160] "PB(L2)"            "GOR(L)"            "EPA(R)"           
#> [163] "INP"               "AL-DL2v(R)"        "AL-VC1(R)"        
#> [166] "AL-DM2(L)"         "AL-VM2(R)"         "AL-DM5(L)"        
#> [169] "FBl1"              "ICL(R)"            "a'3(R)"           
#> [172] "mALT(L)"           "g3(R)"             "AL-DM3(L)"        
#> [175] "AL-DL5(R)"         "AL-VM5d(R)"        "LAL(L)"           
#> [178] "PB(L3)"            "FBl2"              "FBl7"             
#> [181] "AL-VA2(R)"         "AL-VM1(R)"         "AL-VP1d(R)"       
#> [184] "GA(R)"             "AL-DM4(R)"         "NO2(R)"           
#> [187] "AL-VA1v(R)"        "PB(L5)"            "CX"               
#> [190] "hemibrain"         "VMNP"              "a'2(R)"           
#> [193] "ROB(R)"            "AL-DM3(R)"         "mALT(R)"          
#> [196] "NO3(R)"            "FB-column3"        "AL-DC3(R)"        
#> [199] "AL-DA3(L)"         "FBl8"              "LX(L)"            
#> [202] "LAL(R)"            "AL-VA1d(R)"        "AL-VC3m(R)"       
#> [205] "PB(R6)"            "lACA(R)"           "AL-VM7v(L)"       
#> [208] "NO3(L)"            "AL-VC5(R)"         "AL-VM3(R)"        
#> [211] "SCL(R)"            "AL-VM7d(L)"        "AL-VM7d(R)"       
#> [214] "FBl4"              "FBl5"              "AL-DP1l(R)"       
#> [217] "CA(L)"             "BU(L)"             "g4(R)"            
#> [220] "AL-DP1m(R)"        "CRE(-ROB,-RUB)(R)" "CRE(L)"           
#> [223] "AL-DA4m(R)"        "AL-VL2a(R)"        "NO1(L)"           
#> [226] "AL-VA7m(R)"        "AOT(R)"            "b'1(R)"           
#> [229] "SCL(L)"            "g2(R)"             "SNP(R)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "NO"                "EBr3am"            "SIP(L)"           
#>   [4] "FB"                "PLP(R)"            "bL(L)"            
#>   [7] "EBr3pw"            "lACA(R)"           "CRE(R)"           
#>  [10] "FBl4"              "FBl6"              "b1(R)"            
#>  [13] "PB(R4)"            "GA(R)"             "PED(R)"           
#>  [16] "NO3(L)"            "vACA(R)"           "a2(R)"            
#>  [19] "PB(R9)"            "FBl7"              "VES(L)"           
#>  [22] "AL(R)"             "VLNP(R)"           "SPS(R)"           
#>  [25] "PB(R5)"            "FBl2"              "EBr1"             
#>  [28] "PB(L6)"            "PB(R1)"            "CRE(L)"           
#>  [31] "SMP(R)"            "GOR(L)"            "hemibrain"        
#>  [34] "GC"                "g1(R)"             "ICL(L)"           
#>  [37] "FBl5"              "b'L(L)"            "OL(R)"            
#>  [40] "g3(R)"             "LH(R)"             "ATL(R)"           
#>  [43] "PB(L4)"            "IB"                "ATL(L)"           
#>  [46] "CA(L)"             "NO2(L)"            "PB(L3)"           
#>  [49] "PB(R3)"            "POC"               "ME(R)"            
#>  [52] "MB(L)"             "NO1(R)"            "aL(L)"            
#>  [55] "SCL(L)"            "mALT(R)"           "SPS(L)"           
#>  [58] "VMNP"              "a'L(R)"            "FBl3"             
#>  [61] "FBl9"              "CA(R)"             "ICL(R)"           
#>  [64] "WED(R)"            "CX"                "FBl1"             
#>  [67] "PB(R8)"            "INP"               "LOP(R)"           
#>  [70] "mALT(L)"           "gL(R)"             "NO(R)"            
#>  [73] "CAN(R)"            "bL(R)"             "a'3(R)"           
#>  [76] "gL(L)"             "BU(L)"             "PB(R2)"           
#>  [79] "aL(R)"             "PB(L7)"            "PB(L2)"           
#>  [82] "PB(L1)"            "b'L(R)"            "EBr6"             
#>  [85] "LX(R)"             "IPS(R)"            "PENP"             
#>  [88] "g5(R)"             "GNG"               "NO(L)"            
#>  [91] "LAL(R)"            "PRW"               "SCL(R)"           
#>  [94] "SNP(R)"            "MB(+ACA)(R)"       "LX(L)"            
#>  [97] "b'1(R)"            "b2(R)"             "GOR(R)"           
#> [100] "AVLP(R)"           "AOTU(R)"           "PB(R7)"           
#> [103] "a'2(R)"            "AL(L)"             "SIP(R)"           
#> [106] "LO(R)"             "VES(R)"            "GF(R)"            
#> [109] "RUB(L)"            "NO2(R)"            "SLP(R)"           
#> [112] "CRE(-ROB,-RUB)(R)" "SAD(-AMMC)"        "SMP(L)"           
#> [115] "dACA(R)"           "BU(R)"             "a'L(L)"           
#> [118] "MB(R)"             "b'2(R)"            "NO3(R)"           
#> [121] "EBr5"              "EPA(R)"            "RUB(R)"           
#> [124] "PB(L5)"            "SAD"               "g2(R)"            
#> [127] "EBr2r4"            "LAL(L)"            "AME(R)"           
#> [130] "a1(R)"             "FBl8"              "PVLP(R)"          
#> [133] "AB(R)"             "FLA(R)"            "g4(R)"            
#> [136] "ROB(R)"            "a3(R)"             "PB(L8)"           
#> [139] "NO1(L)"            "PB"                "EB"               
#> [142] "EPA(L)"            "AOT(R)"            "LAL(-GA)(R)"      
#> [145] "EBr3d"             "a'1(R)"            "AMMC"             
#> [148] "PB(R6)"            "SNP(L)"            "AB(L)"            
#> [151] "PB(L9)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DM3(L)"         "AL-VA7l(R)"        "AL-VP5(R)"        
#>   [4] "a1(R)"             "b'L(R)"            "FBl9"             
#>   [7] "AL-DL4(R)"         "a2(R)"             "SIP(L)"           
#>  [10] "a'2(R)"            "AL-VA7m(R)"        "AL-VM1(R)"        
#>  [13] "AL-VM4(R)"         "PB(R6)"            "AL-VA5(R)"        
#>  [16] "AL-V(R)"           "mALT(L)"           "AL-DL5(L)"        
#>  [19] "AL-DM1(L)"         "PVLP(R)"           "FBl5"             
#>  [22] "bL(R)"             "AL-VP4(R)"         "AL(L)"            
#>  [25] "ROB(R)"            "AL-VL2p(R)"        "AL-DA3(R)"        
#>  [28] "AL-DM6(R)"         "AL-VP1d(R)"        "AB(L)"            
#>  [31] "EBr5"              "SPS(L)"            "FBl3"             
#>  [34] "LAL(L)"            "EBr2r4"            "EBr3pw"           
#>  [37] "FBl2"              "AL-DL1(R)"         "AL-DP1m(R)"       
#>  [40] "vACA(R)"           "FBl4"              "PB(R7)"           
#>  [43] "PB(R3)"            "SMP(L)"            "GOR(L)"           
#>  [46] "FBl1"              "b1(R)"             "AL-DM5(R)"        
#>  [49] "SCL(R)"            "PB(L8)"            "NO1(R)"           
#>  [52] "PB(L6)"            "VES(R)"            "AL-VC3l(R)"       
#>  [55] "PLP(R)"            "SNP(R)"            "EPA(R)"           
#>  [58] "AL-DL4(L)"         "LO(R)"             "LH(R)"            
#>  [61] "AB(R)"             "AL-DM3(R)"         "AL-VA2(R)"        
#>  [64] "AL-VP2(R)"         "dACA(R)"           "PB(R9)"           
#>  [67] "MB(+ACA)(R)"       "EPA(L)"            "MB(R)"            
#>  [70] "AL-DL5(R)"         "g4(R)"             "FB-column3"       
#>  [73] "AL(R)"             "b'1(R)"            "AL-D(L)"          
#>  [76] "GC"                "FB"                "SIP(R)"           
#>  [79] "CRE(L)"            "aL(L)"             "AL-VA1v(R)"       
#>  [82] "IB"                "AMMC"              "MB(L)"            
#>  [85] "AL-DA4m(L)"        "FBl7"              "AL-VL1(R)"        
#>  [88] "PED(R)"            "g2(R)"             "AL-D(R)"          
#>  [91] "AL-VA6(R)"         "AL-VL2a(R)"        "AL-VP1m(R)"       
#>  [94] "GA(R)"             "PB(L4)"            "AL-DA1(R)"        
#>  [97] "AL-VC5(R)"         "AL-VM7v(L)"        "AL-VM7d(R)"       
#> [100] "AL-DP1m(L)"        "a'1(R)"            "AL-DC4(L)"        
#> [103] "PB(R4)"            "AL-DC3(R)"         "LOP(R)"           
#> [106] "CX"                "PB(R1)"            "INP"              
#> [109] "RUB(L)"            "OL(R)"             "AL-DC1(R)"        
#> [112] "PB(L3)"            "WED(R)"            "AL-VA6(L)"        
#> [115] "FBl6"              "LAL(R)"            "RUB(R)"           
#> [118] "CA(R)"             "NO2(R)"            "a'L(L)"           
#> [121] "AL-VC2(R)"         "SMP(R)"            "CRE(-ROB,-RUB)(R)"
#> [124] "NO(L)"             "aL(R)"             "CAN(R)"           
#> [127] "AL-VM2(R)"         "b'2(R)"            "AL-DC2(L)"        
#> [130] "PB(L2)"            "g3(R)"             "NO1(L)"           
#> [133] "PB(L1)"            "GOR(R)"            "AL-DA4m(R)"       
#> [136] "AL-DC4(R)"         "AL-DL2v(R)"        "EBr3am"           
#> [139] "PRW"               "g1(R)"             "b'L(L)"           
#> [142] "lACA(R)"           "BU(R)"             "AL-VM7d(L)"       
#> [145] "PB(R2)"            "AL-VA1d(R)"        "AL-VA3(R)"        
#> [148] "AL-VA4(R)"         "AL-VM7v(R)"        "ME(R)"            
#> [151] "CRE(-RUB)(L)"      "AL-VP3(R)"         "CRE(R)"           
#> [154] "VMNP"              "ATL(R)"            "GNG"              
#> [157] "AL-VC1(R)"         "AL-VM3(R)"         "LX(R)"            
#> [160] "SNP(L)"            "AL-DM4(R)"         "AL-VM5v(R)"       
#> [163] "AL-DA2(L)"         "PB(L9)"            "AL-DM2(R)"        
#> [166] "GF(R)"             "EBr6"              "a'L(R)"           
#> [169] "AL-DA2(R)"         "AL-VC4(R)"         "FLA(R)"           
#> [172] "AL-DC2(R)"         "hemibrain"         "EBr3d"            
#> [175] "a'3(R)"            "PB(L5)"            "PENP"             
#> [178] "AL-DM5(L)"         "AL-DM2(L)"         "AL-DA4l(R)"       
#> [181] "AL-DP1l(R)"        "AOTU(R)"           "AL-DM4(L)"        
#> [184] "IPS(R)"            "AL-DL3(R)"         "CA(L)"            
#> [187] "NO2(L)"            "PB(R8)"            "PB"               
#> [190] "ICL(R)"            "AL-DM1(R)"         "AL-VC3m(R)"       
#> [193] "ATL(L)"            "EBr1"              "BU(L)"            
#> [196] "AL-DC1(L)"         "NO(R)"             "NO3(L)"           
#> [199] "VES(L)"            "VLNP(R)"           "SAD"              
#> [202] "LAL(-GA)(R)"       "PB(L7)"            "AL-VP1l(R)"       
#> [205] "EB"                "bL(L)"             "FBl8"             
#> [208] "LX(L)"             "SCL(L)"            "gL(R)"            
#> [211] "ICL(L)"            "AVLP(R)"           "AME(R)"           
#> [214] "a3(R)"             "AOT(R)"            "b2(R)"            
#> [217] "AL-DM6(L)"         "SAD(-AMMC)"        "PB(R5)"           
#> [220] "SPS(R)"            "mALT(R)"           "NO"               
#> [223] "gL(L)"             "AL-VM5d(R)"        "NO3(R)"           
#> [226] "AL-DA3(L)"         "g5(R)"             "POC"              
#> [229] "AL-DL2d(R)"        "SLP(R)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-DC2(L)"         "AVLP(R)"           "b'L(L)"           
#>   [4] "VES(L)"            "AL-VM7d(R)"        "AL-VM5v(R)"       
#>   [7] "NO2(L)"            "SMP(L)"            "aL(R)"            
#>  [10] "FLA(R)"            "AL-DL5(L)"         "VES(R)"           
#>  [13] "ATL(R)"            "PED(R)"            "dACA(R)"          
#>  [16] "CA(R)"             "AL-DL1(R)"         "EPA(L)"           
#>  [19] "AL-VA2(R)"         "vACA(R)"           "a3(R)"            
#>  [22] "AL-DM2(L)"         "PRW"               "IB"               
#>  [25] "AL-DM4(L)"         "FBl7"              "b'1(R)"           
#>  [28] "EBr2r4"            "PB(R2)"            "AMMC"             
#>  [31] "PB(R1)"            "MB(R)"             "AL-DL2v(R)"       
#>  [34] "SPS(L)"            "AL-DM5(L)"         "AOT(R)"           
#>  [37] "GA(R)"             "g4(R)"             "ICL(R)"           
#>  [40] "GOR(L)"            "ICL(L)"            "aL(L)"            
#>  [43] "AL-DM2(R)"         "CRE(-ROB,-RUB)(R)" "PB(R9)"           
#>  [46] "AL-DL3(R)"         "LAL(-GA)(R)"       "NO(L)"            
#>  [49] "AL-VA1d(R)"        "RUB(R)"            "g1(R)"            
#>  [52] "ROB(R)"            "PB(L1)"            "PB"               
#>  [55] "RUB(L)"            "GNG"               "AL-VA3(R)"        
#>  [58] "AL-VC3l(R)"        "LAL(L)"            "b2(R)"            
#>  [61] "FB-column3"        "AL-DC4(R)"         "LOP(R)"           
#>  [64] "lACA(R)"           "AL(R)"             "AL-DM4(R)"        
#>  [67] "VLNP(R)"           "hemibrain"         "SCL(L)"           
#>  [70] "EBr6"              "AL-DM1(L)"         "bL(R)"            
#>  [73] "AL-VP1d(R)"        "NO3(L)"            "NO2(R)"           
#>  [76] "AL-VC5(R)"         "AL-VP2(R)"         "PB(R6)"           
#>  [79] "a2(R)"             "AL-VM7v(R)"        "PB(L4)"           
#>  [82] "AL-VC3m(R)"        "mALT(R)"           "GC"               
#>  [85] "AL-DM6(L)"         "AL-DC2(R)"         "AL-DM1(R)"        
#>  [88] "AL-DA2(L)"         "LX(L)"             "AL-VP4(R)"        
#>  [91] "AL-VM7d(L)"        "AL-DC1(R)"         "b'2(R)"           
#>  [94] "PB(L6)"            "SNP(L)"            "PVLP(R)"          
#>  [97] "AL-DP1m(L)"        "PB(L7)"            "AL-VC4(R)"        
#> [100] "AL-VL1(R)"         "EBr5"              "AL-VA7m(R)"       
#> [103] "AL-DA3(L)"         "CRE(L)"            "AL-D(R)"          
#> [106] "AL-DM5(R)"         "CX"                "FB"               
#> [109] "SAD(-AMMC)"        "b1(R)"             "INP"              
#> [112] "AL-DA4m(R)"        "AL-DL5(R)"         "FBl6"             
#> [115] "FBl8"              "AL-DP1m(R)"        "AL-VA5(R)"        
#> [118] "AL-VL2p(R)"        "PB(L3)"            "LH(R)"            
#> [121] "AL-DL4(R)"         "GF(R)"             "PB(L9)"           
#> [124] "FBl9"              "LAL(R)"            "AL-VM5d(R)"       
#> [127] "AL-V(R)"           "AL-VA7l(R)"        "g3(R)"            
#> [130] "AOTU(R)"           "NO1(R)"            "a'1(R)"           
#> [133] "AB(R)"             "AL-DP1l(R)"        "AL-VA1v(R)"       
#> [136] "AL-VA4(R)"         "WED(R)"            "CRE(R)"           
#> [139] "FBl3"              "CAN(R)"            "CA(L)"            
#> [142] "NO1(L)"            "g2(R)"             "AL-VM7v(L)"       
#> [145] "g5(R)"             "NO(R)"             "EBr3pw"           
#> [148] "a1(R)"             "EBr1"              "VMNP"             
#> [151] "AL-VA6(L)"         "PB(R3)"            "AL-VC2(R)"        
#> [154] "AL-VA6(R)"         "AL-VM4(R)"         "PENP"             
#> [157] "PB(R8)"            "a'L(R)"            "gL(R)"            
#> [160] "FBl1"              "a'L(L)"            "AME(R)"           
#> [163] "SLP(R)"            "gL(L)"             "CRE(-RUB)(L)"     
#> [166] "AL-VL2a(R)"        "AL-VM3(R)"         "AL-DC1(L)"        
#> [169] "NO"                "FBl5"              "NO3(R)"           
#> [172] "ME(R)"             "AL-VP5(R)"         "POC"              
#> [175] "OL(R)"             "AL-VM1(R)"         "AL-VP3(R)"        
#> [178] "mALT(L)"           "SNP(R)"            "SIP(R)"           
#> [181] "bL(L)"             "EBr3am"            "SIP(L)"           
#> [184] "LX(R)"             "EBr3d"             "EB"               
#> [187] "PB(R7)"            "AL(L)"             "FBl4"             
#> [190] "AL-DM3(R)"         "PB(L8)"            "AL-DA4m(L)"       
#> [193] "AL-VC1(R)"         "AL-DC4(L)"         "LO(R)"            
#> [196] "BU(R)"             "GOR(R)"            "MB(L)"            
#> [199] "b'L(R)"            "PB(R4)"            "AL-DA2(R)"        
#> [202] "AL-DA3(R)"         "a'3(R)"            "a'2(R)"           
#> [205] "AL-DA4l(R)"        "AL-DL2d(R)"        "PB(L2)"           
#> [208] "PB(L5)"            "SPS(R)"            "BU(L)"            
#> [211] "AL-DL4(L)"         "AL-DM3(L)"         "AB(L)"            
#> [214] "SMP(R)"            "SCL(R)"            "EPA(R)"           
#> [217] "AL-D(L)"           "AL-DC3(R)"         "AL-DA1(R)"        
#> [220] "AL-VM2(R)"         "PLP(R)"            "IPS(R)"           
#> [223] "AL-VP1m(R)"        "PB(R5)"            "SAD"              
#> [226] "MB(+ACA)(R)"       "ATL(L)"            "FBl2"             
#> [229] "AL-DM6(R)"         "AL-VP1l(R)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                           
#> [1,] "GET"  "/oauth2callback"                              
#> [2,] "POST" "/api/v:ver/npexplorer/celltype/:dataset/:type"
#> [3,] "GET"  "/api/dbmeta/instances"                        
#> [4,] "GET"  "/api/npexplorer/neuronmetavals"               
#> [5,] "GET"  "/api/v:ver/dbmeta/version"                    
#> [6,] "GET"  "/api/v:ver/dbmeta/database"                   
#>      name                                                                                          
#> [1,] "github.com/janelia-flyem/echo-secure.oauthCallbackHandler"                                   
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDataInstances-fm"         
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMetaVals-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }