Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "EPA(L)"            "AL-VP1l(R)"        "CRE(L)"           
#>   [4] "AL-VL1(R)"         "a'2(R)"            "AL-DM1(L)"        
#>   [7] "CA(R)"             "bL(R)"             "PENP"             
#>  [10] "FBl6"              "AL(L)"             "PB(L6)"           
#>  [13] "FB-column3"        "FBl2"              "AL-DL4(R)"        
#>  [16] "AL-VA2(R)"         "lACA(R)"           "SPS(L)"           
#>  [19] "AL-VM4(R)"         "EBr3pw"            "PB(R3)"           
#>  [22] "PB(R5)"            "a'L(L)"            "LAL(-GA)(R)"      
#>  [25] "PB(L7)"            "b'L(L)"            "AL-VA4(R)"        
#>  [28] "b'1(R)"            "AME(R)"            "RUB(L)"           
#>  [31] "vACA(R)"           "a'L(R)"            "INP"              
#>  [34] "b2(R)"             "AL-VA5(R)"         "AL-VC2(R)"        
#>  [37] "RUB(R)"            "AMMC"              "AL-DM4(L)"        
#>  [40] "ATL(R)"            "FBl1"              "LAL(R)"           
#>  [43] "AL-D(R)"           "AL-VC3m(R)"        "ME(R)"            
#>  [46] "EBr2r4"            "PB(L3)"            "EPA(R)"           
#>  [49] "AL-VC5(R)"         "AL-VM5d(R)"        "AL-VM7v(R)"       
#>  [52] "AL(R)"             "AL-VA3(R)"         "CA(L)"            
#>  [55] "AL-DC2(L)"         "AL-DP1m(R)"        "AL-DM2(R)"        
#>  [58] "AL-VL2a(R)"        "g3(R)"             "LX(L)"            
#>  [61] "SAD(-AMMC)"        "NO1(L)"            "bL(L)"            
#>  [64] "PB(L4)"            "AL-DL2d(R)"        "POC"              
#>  [67] "PED(R)"            "PB(L8)"            "AL-DL3(R)"        
#>  [70] "IB"                "ROB(R)"            "GOR(R)"           
#>  [73] "AL-VP1m(R)"        "SIP(L)"            "gL(R)"            
#>  [76] "AL-DL1(R)"         "AL-VA6(R)"         "AL-VC1(R)"        
#>  [79] "LX(R)"             "ICL(L)"            "AL-VA7l(R)"       
#>  [82] "CX"                "GC"                "LO(R)"            
#>  [85] "PB(R4)"            "GNG"               "GOR(L)"           
#>  [88] "VMNP"              "IB112"             "VES(R)"           
#>  [91] "FBl7"              "CAN(R)"            "AL-DM3(R)"        
#>  [94] "SCL(L)"            "aL(R)"             "SCL(R)"           
#>  [97] "ATL(L)"            "FBl9"              "AL-DM1(R)"        
#> [100] "AL-VA1d(R)"        "AL-DM6(L)"         "PB(R9)"           
#> [103] "AL-DA3(R)"         "NO"                "VLNP(R)"          
#> [106] "PRW"               "EB"                "NO2(R)"           
#> [109] "AL-DL4(L)"         "AL-DA1(R)"         "AL-VM3(R)"        
#> [112] "SMP(R)"            "GF(R)"             "AL-VA6(L)"        
#> [115] "NO3(L)"            "FBl4"              "SLP(R)"           
#> [118] "AL-VP2(R)"         "PB(L5)"            "AOTU(R)"          
#> [121] "EBr1"              "IPS(R)"            "AL-DA3(L)"        
#> [124] "SNP(L)"            "FBl8"              "OL(R)"            
#> [127] "AL-VP5(R)"         "AL-VM7v(L)"        "MB(+ACA)(R)"      
#> [130] "AL-V(R)"           "b'2(R)"            "aL(L)"            
#> [133] "dACA(R)"           "AL-VM2(R)"         "AL-DA2(R)"        
#> [136] "SPS(R)"            "EBr3d"             "NO1(R)"           
#> [139] "AL-VP4(R)"         "GA(R)"             "WED(R)"           
#> [142] "b1(R)"             "AL-DA4m(R)"        "AL-DM6(R)"        
#> [145] "AL-VL2p(R)"        "PB(L2)"            "FB"               
#> [148] "AL-DM3(L)"         "PB(R1)"            "AB(R)"            
#> [151] "NO2(L)"            "PVLP(R)"           "AL-VA1v(R)"       
#> [154] "g5(R)"             "AL-DC2(R)"         "AL-DM4(R)"        
#> [157] "CRE(-ROB,-RUB)(R)" "EBr6"              "PLP(R)"           
#> [160] "gL(L)"             "BU(R)"             "SIP(R)"           
#> [163] "FBl5"              "AL-VM1(R)"         "PB(R6)"           
#> [166] "CRE(R)"            "AVLP(R)"           "AL-VP1d(R)"       
#> [169] "EBr3am"            "mALT(L)"           "AL-DA4m(L)"       
#> [172] "FBl3"              "PB(R2)"            "g1(R)"            
#> [175] "g2(R)"             "AL-VP3(R)"         "SMP(L)"           
#> [178] "MB(L)"             "FLA(R)"            "CRE(-RUB)(L)"     
#> [181] "VES(L)"            "AL-VC3l(R)"        "PB(L1)"           
#> [184] "AL-DP1m(L)"        "b'L(R)"            "AB(L)"            
#> [187] "a3(R)"             "PB(R8)"            "AL-DL5(L)"        
#> [190] "SNP(R)"            "MB(R)"             "AL-VM5v(R)"       
#> [193] "AL-DL5(R)"         "AL-DM5(R)"         "a1(R)"            
#> [196] "AL-DC1(L)"         "AL-DA4l(R)"        "AL-DC1(R)"        
#> [199] "PB(L9)"            "ICL(R)"            "AL-VA7m(R)"       
#> [202] "SAD"               "AL-DA2(L)"         "AL-DC4(R)"        
#> [205] "AOT(R)"            "mALT(R)"           "AL-D(L)"          
#> [208] "AL-VM7d(L)"        "a'1(R)"            "LH(R)"            
#> [211] "PB"                "AL-DC4(L)"         "AL-DP1l(R)"       
#> [214] "AL-VC4(R)"         "LOP(R)"            "NO3(R)"           
#> [217] "BU(L)"             "AL-DM2(L)"         "AL-DL2v(R)"       
#> [220] "LAL(L)"            "hemibrain"         "AL-DC3(R)"        
#> [223] "a2(R)"             "AL-VM7d(R)"        "g4(R)"            
#> [226] "AL-DM5(L)"         "NO(R)"             "NO(L)"            
#> [229] "a'3(R)"            "PB(R7)"            "EBr5"             
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "FB"                "EBr2r4"            "hemibrain"        
#>   [4] "VMNP"              "EB"                "FBl5"             
#>   [7] "RUB(L)"            "PENP"              "GA(R)"            
#>  [10] "PB(L9)"            "PVLP(R)"           "IPS(R)"           
#>  [13] "a3(R)"             "a'L(L)"            "PLP(R)"           
#>  [16] "SNP(R)"            "CRE(L)"            "VES(L)"           
#>  [19] "FBl4"              "MB(R)"             "VLNP(R)"          
#>  [22] "SAD"               "AL(L)"             "FBl6"             
#>  [25] "bL(R)"             "PB(L2)"            "aL(R)"            
#>  [28] "g5(R)"             "a'3(R)"            "BU(R)"            
#>  [31] "AVLP(R)"           "WED(R)"            "PB(L3)"           
#>  [34] "LAL(-GA)(R)"       "PB(L8)"            "a'L(R)"           
#>  [37] "ROB(R)"            "gL(R)"             "PB(R5)"           
#>  [40] "NO3(R)"            "NO1(L)"            "FBl7"             
#>  [43] "FBl8"              "EPA(L)"            "PED(R)"           
#>  [46] "CRE(R)"            "CRE(-ROB,-RUB)(R)" "NO"               
#>  [49] "ATL(R)"            "NO2(R)"            "SCL(L)"           
#>  [52] "SPS(L)"            "ICL(R)"            "g4(R)"            
#>  [55] "a'1(R)"            "SNP(L)"            "SMP(R)"           
#>  [58] "LAL(L)"            "g2(R)"             "MB(+ACA)(R)"      
#>  [61] "FBl3"              "CAN(R)"            "PB(R4)"           
#>  [64] "GNG"               "NO(L)"             "ICL(L)"           
#>  [67] "VES(R)"            "a'2(R)"            "ME(R)"            
#>  [70] "PB(R3)"            "b'L(L)"            "LAL(R)"           
#>  [73] "AL(R)"             "b'2(R)"            "PB(L1)"           
#>  [76] "b'1(R)"            "EBr5"              "CA(L)"            
#>  [79] "vACA(R)"           "bL(L)"             "CA(R)"            
#>  [82] "PB(R8)"            "PB(R9)"            "SAD(-AMMC)"       
#>  [85] "AB(R)"             "g3(R)"             "a2(R)"            
#>  [88] "EBr6"              "dACA(R)"           "LX(R)"            
#>  [91] "PB(R2)"            "PRW"               "GC"               
#>  [94] "LX(L)"             "GOR(R)"            "OL(R)"            
#>  [97] "EBr1"              "SIP(R)"            "g1(R)"            
#> [100] "EBr3d"             "NO3(L)"            "AB(L)"            
#> [103] "GOR(L)"            "mALT(R)"           "b1(R)"            
#> [106] "aL(L)"             "POC"               "PB(R6)"           
#> [109] "b'L(R)"            "FLA(R)"            "INP"              
#> [112] "LOP(R)"            "MB(L)"             "SCL(R)"           
#> [115] "LO(R)"             "FBl9"              "RUB(R)"           
#> [118] "EBr3am"            "CX"                "NO2(L)"           
#> [121] "PB(R7)"            "PB(L6)"            "FBl1"             
#> [124] "AOT(R)"            "ATL(L)"            "a1(R)"            
#> [127] "lACA(R)"           "EPA(R)"            "LH(R)"            
#> [130] "SIP(L)"            "b2(R)"             "SPS(R)"           
#> [133] "AME(R)"            "IB"                "SMP(L)"           
#> [136] "AOTU(R)"           "NO1(R)"            "PB(L5)"           
#> [139] "AMMC"              "gL(L)"             "FBl2"             
#> [142] "PB(R1)"            "NO(R)"             "EBr3pw"           
#> [145] "PB(L7)"            "SLP(R)"            "mALT(L)"          
#> [148] "GF(R)"             "BU(L)"             "PB"               
#> [151] "PB(L4)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "PENP"              "PB"                "PB(L7)"           
#>   [4] "aL(R)"             "a'L(R)"            "a1(R)"            
#>   [7] "AL-DC2(L)"         "PB(L4)"            "AL-VA7l(R)"       
#>  [10] "SAD"               "OL(R)"             "AL-DC4(R)"        
#>  [13] "AL-VA1v(R)"        "AME(R)"            "RUB(L)"           
#>  [16] "PB(R4)"            "AL-VA1d(R)"        "AL-VA6(R)"        
#>  [19] "AL-VM7v(R)"        "SPS(L)"            "AL-DM3(L)"        
#>  [22] "MB(+ACA)(R)"       "a'L(L)"            "RUB(R)"           
#>  [25] "EBr3am"            "bL(R)"             "SMP(L)"           
#>  [28] "g3(R)"             "SCL(R)"            "EBr5"             
#>  [31] "AL-DL5(L)"         "NO3(L)"            "PB(L8)"           
#>  [34] "ATL(R)"            "AL-VA5(R)"         "b'L(L)"           
#>  [37] "AL-DL5(R)"         "PB(R5)"            "AL-DM5(L)"        
#>  [40] "IPS(R)"            "GOR(L)"            "g4(R)"            
#>  [43] "BU(R)"             "AL-DA4l(R)"        "AL(L)"            
#>  [46] "AL-DC1(L)"         "AL-D(L)"           "a2(R)"            
#>  [49] "dACA(R)"           "SNP(L)"            "AL-DL1(R)"        
#>  [52] "AL-VA7m(R)"        "PB(R9)"            "PVLP(R)"          
#>  [55] "AL-DM4(R)"         "AL-VP3(R)"         "b'2(R)"           
#>  [58] "PB(R7)"            "FBl9"              "MB(R)"            
#>  [61] "AL-VC5(R)"         "SIP(R)"            "AL-DM2(R)"        
#>  [64] "PRW"               "PB(R1)"            "AL-DL3(R)"        
#>  [67] "AL-VP4(R)"         "PB(R8)"            "PLP(R)"           
#>  [70] "ROB(R)"            "CRE(-RUB)(L)"      "AL-DC3(R)"        
#>  [73] "AL-VM7d(R)"        "EB"                "SAD(-AMMC)"       
#>  [76] "AL-VC2(R)"         "AL-VC3l(R)"        "AMMC"             
#>  [79] "CX"                "AL-DA1(R)"         "AL-VL1(R)"        
#>  [82] "AL-VM5v(R)"        "AL-VP2(R)"         "SCL(L)"           
#>  [85] "ATL(L)"            "NO3(R)"            "PB(L9)"           
#>  [88] "VES(L)"            "AL-V(R)"           "a'1(R)"           
#>  [91] "LH(R)"             "NO2(R)"            "AL-VA2(R)"        
#>  [94] "WED(R)"            "PB(L1)"            "ME(R)"            
#>  [97] "EPA(R)"            "SIP(L)"            "EBr3d"            
#> [100] "EBr2r4"            "EPA(L)"            "AL-DA4m(R)"       
#> [103] "FBl8"              "AL-DM1(L)"         "AL-DL2d(R)"       
#> [106] "PED(R)"            "AL-DM2(L)"         "AL-VM7d(L)"       
#> [109] "FB"                "AL-DC4(L)"         "AL-DL4(L)"        
#> [112] "AL-DA2(R)"         "vACA(R)"           "mALT(L)"          
#> [115] "NO(L)"             "AL-DP1l(R)"        "AL-VP5(R)"        
#> [118] "a'3(R)"            "AL-D(R)"           "hemibrain"        
#> [121] "VMNP"              "ICL(R)"            "INP"              
#> [124] "a3(R)"             "SLP(R)"            "AL-DC2(R)"        
#> [127] "AL-DL2v(R)"        "GF(R)"             "NO1(L)"           
#> [130] "CRE(L)"            "IB"                "AOTU(R)"          
#> [133] "g2(R)"             "bL(L)"             "VES(R)"           
#> [136] "FBl6"              "CRE(-ROB,-RUB)(R)" "CA(R)"            
#> [139] "AL-VC3m(R)"        "LX(L)"             "CAN(R)"           
#> [142] "AL-VA4(R)"         "gL(L)"             "AL-DP1m(R)"       
#> [145] "AB(L)"             "AL-VM7v(L)"        "AL-VM2(R)"        
#> [148] "lACA(R)"           "a'2(R)"            "SNP(R)"           
#> [151] "FBl2"              "FBl3"              "AL-DM5(R)"        
#> [154] "AL-DA2(L)"         "AL-VM3(R)"         "PB(R6)"           
#> [157] "AB(R)"             "GNG"               "AOT(R)"           
#> [160] "NO(R)"             "GOR(R)"            "FBl1"             
#> [163] "AL-DM6(R)"         "AL-VM1(R)"         "b'1(R)"           
#> [166] "b2(R)"             "LX(R)"             "aL(L)"            
#> [169] "AL-DA3(R)"         "AL-VC1(R)"         "mALT(R)"          
#> [172] "CRE(R)"            "g5(R)"             "PB(R3)"           
#> [175] "AL-DA4m(L)"        "AL-DC1(R)"         "AL-DP1m(L)"       
#> [178] "AL-VP1d(R)"        "CA(L)"             "GC"               
#> [181] "AL-DA3(L)"         "AL-DM4(L)"         "NO2(L)"           
#> [184] "GA(R)"             "g1(R)"             "gL(R)"            
#> [187] "PB(L2)"            "PB(L6)"            "AL-VM4(R)"        
#> [190] "AL-DM6(L)"         "FB-column3"        "FBl7"             
#> [193] "AL-VL2a(R)"        "PB(L3)"            "FLA(R)"           
#> [196] "NO1(R)"            "LO(R)"             "LAL(R)"           
#> [199] "AL-DM3(R)"         "PB(R2)"            "MB(L)"            
#> [202] "FBl5"              "AL-DL4(R)"         "AL-DM1(R)"        
#> [205] "EBr3pw"            "POC"               "LAL(L)"           
#> [208] "SMP(R)"            "LOP(R)"            "PB(L5)"           
#> [211] "AL-VP1l(R)"        "EBr6"              "BU(L)"            
#> [214] "ICL(L)"            "AVLP(R)"           "AL-VC4(R)"        
#> [217] "AL-VP1m(R)"        "LAL(-GA)(R)"       "NO"               
#> [220] "AL(R)"             "b'L(R)"            "AL-VA6(L)"        
#> [223] "FBl4"              "AL-VA3(R)"         "VLNP(R)"          
#> [226] "SPS(R)"            "AL-VM5d(R)"        "EBr1"             
#> [229] "b1(R)"             "AL-VL2p(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "FBl4"              "CRE(L)"            "g5(R)"            
#>   [4] "AL-VM7d(R)"        "AL-VP1l(R)"        "AB(R)"            
#>   [7] "AL-VM2(R)"         "b'2(R)"            "PRW"              
#>  [10] "AL-VP3(R)"         "PB(L4)"            "AL-VM7v(R)"       
#>  [13] "mALT(L)"           "AL-DA4l(R)"        "AL-V(R)"          
#>  [16] "a'1(R)"            "AL-VA7l(R)"        "LAL(L)"           
#>  [19] "PB(L2)"            "AL-DM4(R)"         "AL-D(R)"          
#>  [22] "SNP(R)"            "gL(L)"             "FBl3"             
#>  [25] "AL-DC1(R)"         "AL-DA4m(L)"        "AL-DA3(R)"        
#>  [28] "AL-VA4(R)"         "AL-VM5v(R)"        "AL-DM3(L)"        
#>  [31] "PB(L6)"            "AL-VA7m(R)"        "CX"               
#>  [34] "AL-DP1m(R)"        "PB(L3)"            "AL-DA3(L)"        
#>  [37] "AL-DM6(L)"         "NO(L)"             "g2(R)"            
#>  [40] "AME(R)"            "AL-VA1d(R)"        "lACA(R)"          
#>  [43] "LAL(-GA)(R)"       "EBr5"              "CAN(R)"           
#>  [46] "bL(R)"             "AL-DL2v(R)"        "AL(L)"            
#>  [49] "AMMC"              "FB"                "AL-DC2(R)"        
#>  [52] "AL-VA6(R)"         "AVLP(R)"           "IPS(R)"           
#>  [55] "aL(L)"             "AL-VC4(R)"         "AL-VP1m(R)"       
#>  [58] "a'3(R)"            "PB(R6)"            "ROB(R)"           
#>  [61] "PB(R3)"            "AL-VM3(R)"         "VMNP"             
#>  [64] "PED(R)"            "a'2(R)"            "NO3(L)"           
#>  [67] "CRE(-RUB)(L)"      "FBl6"              "PB(R2)"           
#>  [70] "ICL(L)"            "AL-VC3l(R)"        "LAL(R)"           
#>  [73] "AL-VL1(R)"         "AL-VA1v(R)"        "AOT(R)"           
#>  [76] "b'L(R)"            "SNP(L)"            "AL-DC3(R)"        
#>  [79] "AL-DL1(R)"         "g1(R)"             "PVLP(R)"          
#>  [82] "AL-VP2(R)"         "AL-VA3(R)"         "LX(R)"            
#>  [85] "PB(L7)"            "a3(R)"             "AL(R)"            
#>  [88] "SCL(R)"            "BU(R)"             "SMP(R)"           
#>  [91] "GOR(L)"            "NO3(R)"            "a'L(R)"           
#>  [94] "LO(R)"             "SAD(-AMMC)"        "PB(R5)"           
#>  [97] "EBr6"              "AL-VA6(L)"         "AL-VM7d(L)"       
#> [100] "PB(L9)"            "NO2(R)"            "AL-DA1(R)"        
#> [103] "AOTU(R)"           "GA(R)"             "NO"               
#> [106] "CA(L)"             "MB(R)"             "RUB(R)"           
#> [109] "PB"                "INP"               "gL(R)"            
#> [112] "FBl2"              "SCL(L)"            "ME(R)"            
#> [115] "EPA(R)"            "bL(L)"             "SLP(R)"           
#> [118] "POC"               "GF(R)"             "AL-DA2(L)"        
#> [121] "AL-DC1(L)"         "VES(L)"            "AL-VM1(R)"        
#> [124] "IB"                "a2(R)"             "AL-DM4(L)"        
#> [127] "AL-DM3(R)"         "AL-DM6(R)"         "EBr3am"           
#> [130] "ATL(R)"            "AL-DL5(R)"         "AL-VC3m(R)"       
#> [133] "VLNP(R)"           "AL-DC4(L)"         "AL-VC2(R)"        
#> [136] "AL-VC5(R)"         "AL-VP4(R)"         "FB-column3"       
#> [139] "FBl5"              "PB(R4)"            "AL-DL2d(R)"       
#> [142] "WED(R)"            "EBr2r4"            "AL-DP1m(L)"       
#> [145] "NO(R)"             "a1(R)"             "hemibrain"        
#> [148] "SPS(R)"            "FBl1"              "AL-VL2a(R)"       
#> [151] "PB(R7)"            "b'1(R)"            "PB(L5)"           
#> [154] "CRE(-ROB,-RUB)(R)" "CRE(R)"            "AL-DC2(L)"        
#> [157] "PB(R1)"            "AL-VM4(R)"         "ICL(R)"           
#> [160] "GOR(R)"            "AL-DL4(R)"         "vACA(R)"          
#> [163] "EBr1"              "NO1(R)"            "PB(R8)"           
#> [166] "EB"                "b2(R)"             "PB(R9)"           
#> [169] "OL(R)"             "ATL(L)"            "AL-DL4(L)"        
#> [172] "VES(R)"            "FBl8"              "AL-VA2(R)"        
#> [175] "AL-DM5(L)"         "LX(L)"             "AL-VC1(R)"        
#> [178] "AB(L)"             "SIP(L)"            "AL-VM7v(L)"       
#> [181] "MB(+ACA)(R)"       "AL-VP1d(R)"        "GNG"              
#> [184] "BU(L)"             "FLA(R)"            "EBr3d"            
#> [187] "EBr3pw"            "b1(R)"             "AL-DL3(R)"        
#> [190] "AL-VM5d(R)"        "AL-VP5(R)"         "SAD"              
#> [193] "mALT(R)"           "SIP(R)"            "FBl9"             
#> [196] "LOP(R)"            "NO2(L)"            "AL-DA2(R)"        
#> [199] "AL-DP1l(R)"        "g3(R)"             "PB(L1)"           
#> [202] "SPS(L)"            "AL-D(L)"           "PLP(R)"           
#> [205] "a'L(L)"            "NO1(L)"            "g4(R)"            
#> [208] "AL-VA5(R)"         "AL-DM2(R)"         "MB(L)"            
#> [211] "PB(L8)"            "FBl7"              "b'L(L)"           
#> [214] "SMP(L)"            "AL-DA4m(R)"        "AL-DL5(L)"        
#> [217] "CA(R)"             "RUB(L)"            "dACA(R)"          
#> [220] "AL-DM2(L)"         "AL-DM1(R)"         "AL-VL2p(R)"       
#> [223] "PENP"              "GC"                "EPA(L)"           
#> [226] "AL-DM5(R)"         "AL-DM1(L)"         "LH(R)"            
#> [229] "aL(R)"             "AL-DC4(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                               
#> [1,] "HEAD" "/api/*"                           
#> [2,] "GET"  "/api/v:ver/cached/roicompleteness"
#> [3,] "POST" "/api/npexplorer/roiconnectivity"  
#> [4,] "GET"  "/api/npexplorer/simpleconnections"
#> [5,] "POST" "/api/v:ver/raw/cypher/cypher"     
#> [6,] "GET"  "/api/dbmeta/datasets"             
#>      name                                                                                             
#> [1,] "github.com/labstack/echo.glob..func1"                                                           
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIConnectivity-fm"  
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatasets-fm"                 
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }