Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-VA5(R)"         "AL-VM1(R)"         "SAD"              
#>   [4] "AL-VM3(R)"         "EBr3am"            "FBl1"             
#>   [7] "AVLP(R)"           "NO2(R)"            "bL(R)"            
#>  [10] "AL-DL2d(R)"        "CRE(-RUB)(L)"      "AL-DL1(R)"        
#>  [13] "AL-DL2v(R)"        "aL(L)"             "g5(R)"            
#>  [16] "NO2(L)"            "NO1(L)"            "AL-DC3(R)"        
#>  [19] "AL-VC3m(R)"        "IB"                "IPS(R)"           
#>  [22] "a2(R)"             "SNP(R)"            "NO3(L)"           
#>  [25] "AL-VA3(R)"         "AL-VA7m(R)"        "AL-VM5v(R)"       
#>  [28] "PB(R6)"            "AL-DM4(L)"         "AL-VP1l(R)"       
#>  [31] "EBr5"              "EBr1"              "VLNP(R)"          
#>  [34] "GOR(L)"            "SIP(L)"            "MB(L)"            
#>  [37] "SCL(R)"            "AL-D(R)"           "SMP(L)"           
#>  [40] "ICL(R)"            "SLP(R)"            "AL-VM7v(R)"       
#>  [43] "EBr3pw"            "LAL(-GA)(R)"       "LX(R)"            
#>  [46] "EBr2r4"            "PB(R9)"            "NO(L)"            
#>  [49] "g3(R)"             "PB(L4)"            "RUB(R)"           
#>  [52] "CA(L)"             "AL-DP1m(L)"        "AL-DM1(R)"        
#>  [55] "AL-VC2(R)"         "a1(R)"             "AL-DM2(L)"        
#>  [58] "PB(R4)"            "SCL(L)"            "LH(R)"            
#>  [61] "AL-VA6(R)"         "AL-VP1d(R)"        "PB(L5)"           
#>  [64] "AL-DA4l(R)"        "mALT(R)"           "AL-VA1v(R)"       
#>  [67] "a'3(R)"            "PENP"              "b2(R)"            
#>  [70] "AL-VM7v(L)"        "AB(R)"             "g1(R)"            
#>  [73] "CAN(R)"            "AL-DM3(R)"         "AL-VL2p(R)"       
#>  [76] "FBl8"              "WED(R)"            "EBr6"             
#>  [79] "PB(L9)"            "CRE(L)"            "PB(L7)"           
#>  [82] "IB112"             "PVLP(R)"           "PB(L3)"           
#>  [85] "vACA(R)"           "AL-DM2(R)"         "AL-DM4(R)"        
#>  [88] "AL-VL2a(R)"        "AL-VP1m(R)"        "lACA(R)"          
#>  [91] "AL-DL3(R)"         "FBl3"              "FBl2"             
#>  [94] "AL(R)"             "ME(R)"             "PB(R1)"           
#>  [97] "AL-VC1(R)"         "PB(R5)"            "PB(R7)"           
#> [100] "gL(R)"             "ATL(R)"            "POC"              
#> [103] "ATL(L)"            "b'L(R)"            "AL-VA6(L)"        
#> [106] "FBl5"              "FBl9"              "AL-VM4(R)"        
#> [109] "CX"                "a'2(R)"            "AL-VM7d(R)"       
#> [112] "aL(R)"             "AME(R)"            "AL-DA2(R)"        
#> [115] "a'1(R)"            "GC"                "AL-DM6(R)"        
#> [118] "MB(R)"             "MB(+ACA)(R)"       "FLA(R)"           
#> [121] "AL-DP1l(R)"        "AL-VM2(R)"         "FB"               
#> [124] "AL-DL5(L)"         "bL(L)"             "CA(R)"            
#> [127] "a3(R)"             "AL-DM5(R)"         "GA(R)"            
#> [130] "LX(L)"             "LAL(R)"            "AL-VM7d(L)"       
#> [133] "INP"               "GNG"               "BU(L)"            
#> [136] "FBl7"              "EB"                "ROB(R)"           
#> [139] "b'2(R)"            "NO(R)"             "EBr3d"            
#> [142] "a'L(L)"            "PED(R)"            "PB(L2)"           
#> [145] "LOP(R)"            "CRE(R)"            "AL-DA3(L)"        
#> [148] "AL-DC2(L)"         "AL-VC3l(R)"        "EPA(R)"           
#> [151] "PB(L6)"            "PB(R2)"            "AOT(R)"           
#> [154] "PRW"               "gL(L)"             "FB-column3"       
#> [157] "EPA(L)"            "AL-DA4m(R)"        "b'1(R)"           
#> [160] "PB(L1)"            "VMNP"              "a'L(R)"           
#> [163] "AL-VA7l(R)"        "dACA(R)"           "AL-DC2(R)"        
#> [166] "AL-V(R)"           "GF(R)"             "SAD(-AMMC)"       
#> [169] "AL-DC4(R)"         "AL-DA4m(L)"        "g2(R)"            
#> [172] "AL-D(L)"           "LO(R)"             "AL-DC1(R)"        
#> [175] "SMP(R)"            "hemibrain"         "AL-DA3(R)"        
#> [178] "PB(R3)"            "AL-VP3(R)"         "AL-DC4(L)"        
#> [181] "AL-VP2(R)"         "FBl4"              "AL-VL1(R)"        
#> [184] "AL-DM3(L)"         "b1(R)"             "RUB(L)"           
#> [187] "AL-VA2(R)"         "AL-VP5(R)"         "AL-VP4(R)"        
#> [190] "mALT(L)"           "AMMC"              "NO1(R)"           
#> [193] "NO3(R)"            "PLP(R)"            "b'L(L)"           
#> [196] "AL-VC4(R)"         "SPS(R)"            "AL(L)"            
#> [199] "AL-DM5(L)"         "AL-DP1m(R)"        "PB(R8)"           
#> [202] "PB"                "BU(R)"             "VES(R)"           
#> [205] "AOTU(R)"           "AL-DM1(L)"         "AL-DL5(R)"        
#> [208] "AL-VA4(R)"         "CRE(-ROB,-RUB)(R)" "PB(L8)"           
#> [211] "GOR(R)"            "OL(R)"             "SIP(R)"           
#> [214] "VES(L)"            "AL-DM6(L)"         "NO"               
#> [217] "LAL(L)"            "AL-DL4(L)"         "FBl6"             
#> [220] "SPS(L)"            "AL-DC1(L)"         "AL-DL4(R)"        
#> [223] "AB(L)"             "AL-DA2(L)"         "AL-VC5(R)"        
#> [226] "SNP(L)"            "ICL(L)"            "AL-DA1(R)"        
#> [229] "AL-VA1d(R)"        "AL-VM5d(R)"        "g4(R)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(R3)"            "b1(R)"             "bL(R)"            
#>   [4] "PB(L3)"            "PB(L1)"            "ICL(L)"           
#>   [7] "FBl2"              "b'2(R)"            "AB(R)"            
#>  [10] "RUB(L)"            "lACA(R)"           "PB(R1)"           
#>  [13] "FBl9"              "NO2(R)"            "AMMC"             
#>  [16] "a2(R)"             "SAD"               "VMNP"             
#>  [19] "EBr3pw"            "EPA(R)"            "SAD(-AMMC)"       
#>  [22] "CAN(R)"            "GOR(R)"            "AOTU(R)"          
#>  [25] "BU(L)"             "FLA(R)"            "BU(R)"            
#>  [28] "CA(L)"             "PVLP(R)"           "GNG"              
#>  [31] "PB(L5)"            "SPS(L)"            "VES(R)"           
#>  [34] "PB(R4)"            "PB(L4)"            "PB(L7)"           
#>  [37] "SCL(L)"            "a1(R)"             "g2(R)"            
#>  [40] "PB(R9)"            "AOT(R)"            "a'3(R)"           
#>  [43] "INP"               "NO2(L)"            "SNP(L)"           
#>  [46] "LH(R)"             "ATL(R)"            "LX(L)"            
#>  [49] "CRE(R)"            "EPA(L)"            "VLNP(R)"          
#>  [52] "AB(L)"             "SMP(L)"            "EBr1"             
#>  [55] "EBr2r4"            "POC"               "SLP(R)"           
#>  [58] "AL(R)"             "ATL(L)"            "PLP(R)"           
#>  [61] "EBr3d"             "CA(R)"             "PB(R2)"           
#>  [64] "IB"                "PB(R6)"            "PED(R)"           
#>  [67] "MB(L)"             "EB"                "LX(R)"            
#>  [70] "ROB(R)"            "aL(L)"             "g3(R)"            
#>  [73] "hemibrain"         "b'L(R)"            "NO(L)"            
#>  [76] "gL(L)"             "PB(L6)"            "PB(L9)"           
#>  [79] "AME(R)"            "GC"                "b2(R)"            
#>  [82] "FB"                "EBr6"              "RUB(R)"           
#>  [85] "mALT(L)"           "aL(R)"             "SIP(R)"           
#>  [88] "FBl4"              "b'L(L)"            "NO(R)"            
#>  [91] "b'1(R)"            "PRW"               "WED(R)"           
#>  [94] "ME(R)"             "NO"                "NO3(R)"           
#>  [97] "LO(R)"             "FBl1"              "PB(R7)"           
#> [100] "a'L(R)"            "PB(R5)"            "SPS(R)"           
#> [103] "g1(R)"             "FBl3"              "MB(+ACA)(R)"      
#> [106] "FBl5"              "PENP"              "g4(R)"            
#> [109] "gL(R)"             "a'1(R)"            "MB(R)"            
#> [112] "SIP(L)"            "LAL(-GA)(R)"       "a'2(R)"           
#> [115] "SNP(R)"            "LOP(R)"            "PB(L8)"           
#> [118] "g5(R)"             "ICL(R)"            "IPS(R)"           
#> [121] "EBr3am"            "LAL(L)"            "GA(R)"            
#> [124] "PB"                "FBl7"              "AL(L)"            
#> [127] "a3(R)"             "SMP(R)"            "vACA(R)"          
#> [130] "a'L(L)"            "PB(L2)"            "mALT(R)"          
#> [133] "dACA(R)"           "FBl6"              "LAL(R)"           
#> [136] "AVLP(R)"           "GOR(L)"            "NO1(L)"           
#> [139] "CRE(L)"            "NO3(L)"            "GF(R)"            
#> [142] "SCL(R)"            "NO1(R)"            "PB(R8)"           
#> [145] "VES(L)"            "CRE(-ROB,-RUB)(R)" "CX"               
#> [148] "EBr5"              "bL(L)"             "FBl8"             
#> [151] "OL(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "SIP(L)"            "AL-VA2(R)"         "AL-VC3m(R)"       
#>   [4] "NO3(R)"            "g5(R)"             "AL-DC1(R)"        
#>   [7] "AL-DM4(R)"         "RUB(R)"            "PB(L2)"           
#>  [10] "AOTU(R)"           "PB(R7)"            "PB(L6)"           
#>  [13] "OL(R)"             "AL-DA4l(R)"        "AL-DA4m(R)"       
#>  [16] "PLP(R)"            "AL-DM6(L)"         "NO3(L)"           
#>  [19] "FBl7"              "PB(R2)"            "AL-DM3(R)"        
#>  [22] "AB(L)"             "g1(R)"             "BU(L)"            
#>  [25] "a'L(R)"            "a'L(L)"            "AL-D(L)"          
#>  [28] "AMMC"              "a'1(R)"            "MB(+ACA)(R)"      
#>  [31] "AVLP(R)"           "AL-VA5(R)"         "AL-VL2a(R)"       
#>  [34] "AL-DL1(R)"         "LX(R)"             "EBr2r4"           
#>  [37] "FB-column3"        "ICL(R)"            "AL(R)"            
#>  [40] "IPS(R)"            "RUB(L)"            "AL-DA2(R)"        
#>  [43] "NO2(L)"            "PENP"              "GC"               
#>  [46] "EBr1"              "AL-VM2(R)"         "EB"               
#>  [49] "INP"               "a3(R)"             "AL-VL2p(R)"       
#>  [52] "AL-DA3(L)"         "CRE(-RUB)(L)"      "PVLP(R)"          
#>  [55] "b'L(L)"            "AL-VM7d(R)"        "GA(R)"            
#>  [58] "WED(R)"            "EBr6"              "MB(R)"            
#>  [61] "AL-VM3(R)"         "a'3(R)"            "LOP(R)"           
#>  [64] "hemibrain"         "LO(R)"             "CAN(R)"           
#>  [67] "AL-DL2v(R)"        "AL-VM1(R)"         "AL-VP2(R)"        
#>  [70] "AOT(R)"            "mALT(R)"           "FBl6"             
#>  [73] "PED(R)"            "NO1(L)"            "NO1(R)"           
#>  [76] "SNP(R)"            "bL(L)"             "FBl4"             
#>  [79] "LAL(R)"            "AL-VA7l(R)"        "b2(R)"            
#>  [82] "AL-DP1m(L)"        "CA(R)"             "PB(R4)"           
#>  [85] "AL-VA4(R)"         "AL-V(R)"           "PB(L1)"           
#>  [88] "gL(L)"             "AL-DM2(L)"         "NO(L)"            
#>  [91] "EBr3am"            "GF(R)"             "AL-VM7d(L)"       
#>  [94] "AL-DL5(R)"         "AL-VP3(R)"         "AL-VP1l(R)"       
#>  [97] "lACA(R)"           "CRE(L)"            "AL-DP1m(R)"       
#> [100] "AL-DC2(R)"         "AL-DL3(R)"         "AL-VC4(R)"        
#> [103] "AL-VP1m(R)"        "AL-VP1d(R)"        "AL(L)"            
#> [106] "ATL(R)"            "AL-DL4(R)"         "AL-VA1d(R)"       
#> [109] "AL-DL2d(R)"        "AL-VM7v(R)"        "EBr3pw"           
#> [112] "AL-VA7m(R)"        "BU(R)"             "AL-DA4m(L)"       
#> [115] "AL-DM5(R)"         "CX"                "SMP(L)"           
#> [118] "CRE(R)"            "g2(R)"             "AL-DC3(R)"        
#> [121] "AL-VA1v(R)"        "AL-DM6(R)"         "ATL(L)"           
#> [124] "a'2(R)"            "VES(R)"            "FBl2"             
#> [127] "AL-VC2(R)"         "AL-VM7v(L)"        "SIP(R)"           
#> [130] "LAL(L)"            "SPS(L)"            "PB(R8)"           
#> [133] "AL-DL4(L)"         "ROB(R)"            "aL(R)"            
#> [136] "AL-VC1(R)"         "PRW"               "b'L(R)"           
#> [139] "AL-DC1(L)"         "NO"                "AL-VC5(R)"        
#> [142] "AL-VM5d(R)"        "ME(R)"             "AL-DA3(R)"        
#> [145] "PB(R6)"            "POC"               "SMP(R)"           
#> [148] "mALT(L)"           "b'2(R)"            "FBl5"             
#> [151] "AL-D(R)"           "AL-VP4(R)"         "SPS(R)"           
#> [154] "NO2(R)"            "AL-DC4(R)"         "PB(L9)"           
#> [157] "FBl3"              "FBl9"              "NO(R)"            
#> [160] "PB"                "AL-DL5(L)"         "LH(R)"            
#> [163] "GOR(R)"            "SCL(L)"            "a1(R)"            
#> [166] "PB(L3)"            "gL(R)"             "AL-DM3(L)"        
#> [169] "PB(R9)"            "AB(R)"             "AL-DM2(R)"        
#> [172] "AL-VL1(R)"         "VLNP(R)"           "b'1(R)"           
#> [175] "LAL(-GA)(R)"       "AL-DC4(L)"         "IB"               
#> [178] "AME(R)"            "LX(L)"             "SAD"              
#> [181] "EBr5"              "FLA(R)"            "PB(L8)"           
#> [184] "AL-DA2(L)"         "aL(L)"             "AL-DM4(L)"        
#> [187] "AL-DM5(L)"         "PB(L5)"            "dACA(R)"          
#> [190] "ICL(L)"            "MB(L)"             "AL-DP1l(R)"       
#> [193] "PB(R1)"            "bL(R)"             "EBr3d"            
#> [196] "b1(R)"             "AL-DM1(L)"         "AL-VM5v(R)"       
#> [199] "CRE(-ROB,-RUB)(R)" "g3(R)"             "g4(R)"            
#> [202] "EPA(R)"            "PB(L4)"            "SLP(R)"           
#> [205] "AL-VA6(R)"         "CA(L)"             "AL-VA6(L)"        
#> [208] "PB(L7)"            "FBl1"              "FBl8"             
#> [211] "PB(R3)"            "GNG"               "AL-DC2(L)"        
#> [214] "VES(L)"            "GOR(L)"            "FB"               
#> [217] "EPA(L)"            "AL-DA1(R)"         "vACA(R)"          
#> [220] "SNP(L)"            "AL-DM1(R)"         "AL-VA3(R)"        
#> [223] "AL-VP5(R)"         "SCL(R)"            "VMNP"             
#> [226] "SAD(-AMMC)"        "AL-VC3l(R)"        "PB(R5)"           
#> [229] "a2(R)"             "AL-VM4(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "VMNP"              "AL-DC4(R)"         "SPS(L)"           
#>   [4] "FBl5"              "ROB(R)"            "NO(L)"            
#>   [7] "AL-VP1d(R)"        "FBl8"              "AL-VA5(R)"        
#>  [10] "b'1(R)"            "b'2(R)"            "LAL(L)"           
#>  [13] "EBr6"              "NO(R)"             "FBl9"             
#>  [16] "GOR(R)"            "PENP"              "g4(R)"            
#>  [19] "LAL(R)"            "PB(R6)"            "AL-DL5(L)"        
#>  [22] "AL-VC3l(R)"        "ATL(L)"            "a3(R)"            
#>  [25] "AL-DA4m(R)"        "mALT(L)"           "g2(R)"            
#>  [28] "g1(R)"             "g3(R)"             "LH(R)"            
#>  [31] "a'L(L)"            "PB(L3)"            "hemibrain"        
#>  [34] "AL-DP1l(R)"        "PB(R9)"            "CAN(R)"           
#>  [37] "SNP(L)"            "GNG"               "POC"              
#>  [40] "AMMC"              "mALT(R)"           "dACA(R)"          
#>  [43] "AL-DM2(L)"         "FBl3"              "AL-DL3(R)"        
#>  [46] "AL-VM2(R)"         "PB"                "FBl7"             
#>  [49] "VES(L)"            "AB(L)"             "EBr1"             
#>  [52] "a'1(R)"            "bL(L)"             "NO3(L)"           
#>  [55] "OL(R)"             "AL-DM3(R)"         "AL-VA4(R)"        
#>  [58] "AL-DL5(R)"         "AL-VM3(R)"         "GA(R)"            
#>  [61] "AL-VA6(L)"         "AL-VP5(R)"         "SMP(L)"           
#>  [64] "PB(L4)"            "AL-DM4(L)"         "CA(R)"            
#>  [67] "SIP(L)"            "SAD"               "AL-DA1(R)"        
#>  [70] "AL-DC4(L)"         "ICL(L)"            "a2(R)"            
#>  [73] "SAD(-AMMC)"        "AL-VM7v(L)"        "AL-VA7l(R)"       
#>  [76] "AL-VC2(R)"         "AL-V(R)"           "EBr3am"           
#>  [79] "AL-DA2(L)"         "PB(R1)"            "FBl4"             
#>  [82] "INP"               "AL-DM2(R)"         "AL-VP2(R)"        
#>  [85] "a'3(R)"            "AL-DM3(L)"         "b1(R)"            
#>  [88] "AL(R)"             "AL-VA1v(R)"        "AL-VC3m(R)"       
#>  [91] "PB(R2)"            "ICL(R)"            "AL-VA6(R)"        
#>  [94] "AL-VP4(R)"         "PB(L7)"            "ATL(R)"           
#>  [97] "b'L(L)"            "AL-VL1(R)"         "AL-VM5d(R)"       
#> [100] "FBl6"              "CX"                "PED(R)"           
#> [103] "LX(R)"             "PB(L2)"            "PRW"              
#> [106] "SPS(R)"            "PB(R8)"            "aL(L)"            
#> [109] "AL-VA7m(R)"        "LAL(-GA)(R)"       "FLA(R)"           
#> [112] "PVLP(R)"           "a'L(R)"            "RUB(L)"           
#> [115] "AL-DL2v(R)"        "AOTU(R)"           "gL(R)"            
#> [118] "AVLP(R)"           "LX(L)"             "AL-VA2(R)"        
#> [121] "NO2(L)"            "AL-VC4(R)"         "GF(R)"            
#> [124] "BU(L)"             "a'2(R)"            "AL-DL4(L)"        
#> [127] "AL-D(L)"           "AL-DA3(R)"         "AL-VA3(R)"        
#> [130] "AL-VL2a(R)"        "PB(L5)"            "GC"               
#> [133] "SNP(R)"            "AL-DM1(R)"         "VLNP(R)"          
#> [136] "RUB(R)"            "SCL(L)"            "CA(L)"            
#> [139] "AL-DA3(L)"         "FBl1"              "FBl2"             
#> [142] "AL-DP1m(R)"        "AL-VM7v(R)"        "AL-VC5(R)"        
#> [145] "EBr3d"             "AL-VM7d(L)"        "LO(R)"            
#> [148] "MB(+ACA)(R)"       "PB(R3)"            "IPS(R)"           
#> [151] "AL-DL2d(R)"        "AOT(R)"            "EB"               
#> [154] "PB(L9)"            "AL-VA1d(R)"        "SMP(R)"           
#> [157] "LOP(R)"            "AL-DA4m(L)"        "PB(L6)"           
#> [160] "AL-VL2p(R)"        "EBr5"              "BU(R)"            
#> [163] "NO2(R)"            "AL-DA4l(R)"        "AL-DM6(R)"        
#> [166] "AL-VM1(R)"         "FB"                "b'L(R)"           
#> [169] "AL-DM1(L)"         "NO"                "AL-DC1(R)"        
#> [172] "AL-DM5(R)"         "AL-D(R)"           "lACA(R)"          
#> [175] "PLP(R)"            "AL-DP1m(L)"        "PB(R4)"           
#> [178] "MB(R)"             "GOR(L)"            "FB-column3"       
#> [181] "a1(R)"             "b2(R)"             "NO1(R)"           
#> [184] "AL-DC2(L)"         "PB(L1)"            "g5(R)"            
#> [187] "AL-VM5v(R)"        "PB(R5)"            "NO1(L)"           
#> [190] "SCL(R)"            "PB(L8)"            "ME(R)"            
#> [193] "CRE(-RUB)(L)"      "AL-VP3(R)"         "NO3(R)"           
#> [196] "CRE(R)"            "MB(L)"             "PB(R7)"           
#> [199] "gL(L)"             "AL-DC1(L)"         "SIP(R)"           
#> [202] "AME(R)"            "AL-DC3(R)"         "vACA(R)"          
#> [205] "EBr3pw"            "IB"                "AL-DM5(L)"        
#> [208] "AL-DC2(R)"         "CRE(-ROB,-RUB)(R)" "EPA(L)"           
#> [211] "bL(R)"             "AL-DL1(R)"         "AL-VC1(R)"        
#> [214] "AL-DA2(R)"         "AB(R)"             "AL-VP1m(R)"       
#> [217] "WED(R)"            "aL(R)"             "EBr2r4"           
#> [220] "AL(L)"             "AL-DM4(R)"         "AL-DM6(L)"        
#> [223] "SLP(R)"            "AL-VP1l(R)"        "EPA(R)"           
#> [226] "VES(R)"            "AL-DL4(R)"         "AL-VM4(R)"        
#> [229] "AL-VM7d(R)"        "CRE(L)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                        
#> [1,] "GET"  "/api/dbmeta/instances"                     
#> [2,] "GET"  "/api/npexplorer/neuronmetavals"            
#> [3,] "GET"  "/api/v:ver/dbmeta/version"                 
#> [4,] "GET"  "/api/v:ver/dbmeta/database"                
#> [5,] "POST" "/api/v:ver/npexplorer/autapses"            
#> [6,] "GET"  "/api/v:ver/raw/keyvalue/key/:instance/:key"
#>      name                                                                                          
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDataInstances-fm"         
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMetaVals-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }