Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL(L)"             "SIP(R)"            "EBr3pw"           
#>   [4] "mALT(R)"           "a'L(R)"            "CAN(R)"           
#>   [7] "AB(L)"             "PED(R)"            "NO1(L)"           
#>  [10] "FBl8"              "AL-DA1(R)"         "AL-VA1d(R)"       
#>  [13] "AVLP(R)"           "a'L(L)"            "AL-DC4(R)"        
#>  [16] "AL-VA5(R)"         "AL-VM4(R)"         "NO3(L)"           
#>  [19] "NO2(R)"            "vACA(R)"           "AL-DC4(L)"        
#>  [22] "LAL(L)"            "AL-DC1(L)"         "ICL(L)"           
#>  [25] "AL-DL4(R)"         "AL-VA7m(R)"        "AL-VC2(R)"        
#>  [28] "GOR(L)"            "PB(R8)"            "lACA(R)"          
#>  [31] "g1(R)"             "AL-DA4m(R)"        "AL-DM4(R)"        
#>  [34] "AL-VP2(R)"         "SPS(R)"            "AL-DP1m(L)"       
#>  [37] "a'1(R)"            "AOT(R)"            "g4(R)"            
#>  [40] "PB(L3)"            "CRE(R)"            "FLA(R)"           
#>  [43] "SPS(L)"            "FBl5"              "AB(R)"            
#>  [46] "PB(L6)"            "AL-DM3(L)"         "AL-DC3(R)"        
#>  [49] "RUB(R)"            "INP"               "AL-VA7l(R)"       
#>  [52] "a1(R)"             "PLP(R)"            "AL-DA4l(R)"       
#>  [55] "SMP(R)"            "b'1(R)"            "PB(R6)"           
#>  [58] "g3(R)"             "PB(R7)"            "LAL(-GA)(R)"      
#>  [61] "AL-VM7v(L)"        "CA(R)"             "NO3(R)"           
#>  [64] "SNP(R)"            "AL-DL5(R)"         "AL-VA1v(R)"       
#>  [67] "AL-DA4m(L)"        "AL-VP5(R)"         "CA(L)"            
#>  [70] "PB(R1)"            "bL(L)"             "AL-VA4(R)"        
#>  [73] "AL-VM1(R)"         "mALT(L)"           "LH(R)"            
#>  [76] "AL-VC3m(R)"        "PVLP(R)"           "a'3(R)"           
#>  [79] "PB(L5)"            "aL(L)"             "AL-DA2(R)"        
#>  [82] "AL-VA2(R)"         "GF(R)"             "CRE(L)"           
#>  [85] "g5(R)"             "PB(R4)"            "MB(R)"            
#>  [88] "AL-VC3l(R)"        "a3(R)"             "AOTU(R)"          
#>  [91] "PRW"               "AL-VA6(L)"         "PB(L9)"           
#>  [94] "PB(R2)"            "RUB(L)"            "GNG"              
#>  [97] "PB(L2)"            "EBr5"              "AME(R)"           
#> [100] "AL-DL3(R)"         "AL-DM1(R)"         "AL-DC1(R)"        
#> [103] "b'2(R)"            "AL-VP1d(R)"        "POC"              
#> [106] "EBr2r4"            "OL(R)"             "ATL(L)"           
#> [109] "WED(R)"            "ME(R)"             "FBl3"             
#> [112] "AL(R)"             "gL(L)"             "gL(R)"            
#> [115] "AL-D(R)"           "LX(R)"             "AL-D(L)"          
#> [118] "AL-VM5v(R)"        "FBl9"              "PENP"             
#> [121] "BU(R)"             "ROB(R)"            "FBl2"             
#> [124] "SLP(R)"            "AL-VM5d(R)"        "AL-DM2(L)"        
#> [127] "AL-VM2(R)"         "g2(R)"             "CRE(-RUB)(L)"     
#> [130] "AL-VM7d(R)"        "SIP(L)"            "a'2(R)"           
#> [133] "IPS(R)"            "VES(L)"            "AL-DM5(R)"        
#> [136] "a2(R)"             "PB(R9)"            "AL-DM3(R)"        
#> [139] "FBl6"              "b'L(L)"            "AL-DL2v(R)"       
#> [142] "EBr3am"            "PB"                "AL-VM7d(L)"       
#> [145] "GOR(R)"            "AL-DL2d(R)"        "EB"               
#> [148] "AL-VP1m(R)"        "LO(R)"             "FBl7"             
#> [151] "SMP(L)"            "GA(R)"             "aL(R)"            
#> [154] "EBr6"              "AL-DC2(L)"         "AL-DM5(L)"        
#> [157] "PB(R5)"            "b1(R)"             "NO(L)"            
#> [160] "AL-DA3(R)"         "AL-DC2(R)"         "AL-VA6(R)"        
#> [163] "AL-VC4(R)"         "AL-DM4(L)"         "VES(R)"           
#> [166] "AL-VC5(R)"         "AL-VP1l(R)"        "MB(L)"            
#> [169] "BU(L)"             "SCL(L)"            "FBl1"             
#> [172] "bL(R)"             "PB(L8)"            "NO(R)"            
#> [175] "PB(L7)"            "AL-DP1l(R)"        "IB"               
#> [178] "CRE(-ROB,-RUB)(R)" "SAD"               "LX(L)"            
#> [181] "AL-DL1(R)"         "AL-DM2(R)"         "AL-DP1m(R)"       
#> [184] "VLNP(R)"           "NO2(L)"            "b2(R)"            
#> [187] "AL-DM6(L)"         "ATL(R)"            "LAL(R)"           
#> [190] "AL-DM6(R)"         "AL-VL2p(R)"        "VMNP"             
#> [193] "EBr3d"             "AL-DL4(L)"         "AL-DM1(L)"        
#> [196] "AL-VL1(R)"         "AL-VL2a(R)"        "AL-VM7v(R)"       
#> [199] "hemibrain"         "GC"                "AL-DA2(L)"        
#> [202] "AL-DA3(L)"         "IB112"             "AL-VC1(R)"        
#> [205] "AL-V(R)"           "EBr1"              "AL-DL5(L)"        
#> [208] "FB-column3"        "NO1(R)"            "SAD(-AMMC)"       
#> [211] "ICL(R)"            "LOP(R)"            "AMMC"             
#> [214] "CX"                "SNP(L)"            "PB(L4)"           
#> [217] "AL-VA3(R)"         "b'L(R)"            "NO"               
#> [220] "MB(+ACA)(R)"       "PB(R3)"            "AL-VM3(R)"        
#> [223] "AL-VP3(R)"         "PB(L1)"            "EPA(R)"           
#> [226] "FBl4"              "AL-VP4(R)"         "FB"               
#> [229] "SCL(R)"            "dACA(R)"           "EPA(L)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "b'L(R)"            "BU(R)"             "FBl6"             
#>   [4] "PB(R2)"            "CAN(R)"            "PB(L2)"           
#>   [7] "PENP"              "EBr2r4"            "IB"               
#>  [10] "CA(L)"             "g2(R)"             "EBr1"             
#>  [13] "NO1(R)"            "SAD(-AMMC)"        "FB"               
#>  [16] "MB(L)"             "EBr6"              "EPA(R)"           
#>  [19] "ICL(L)"            "FBl9"              "ICL(R)"           
#>  [22] "NO2(L)"            "GF(R)"             "mALT(L)"          
#>  [25] "NO1(L)"            "PVLP(R)"           "MB(+ACA)(R)"      
#>  [28] "FBl7"              "a3(R)"             "LOP(R)"           
#>  [31] "a'L(R)"            "VES(L)"            "PB(L5)"           
#>  [34] "NO3(L)"            "PB(L8)"            "a'1(R)"           
#>  [37] "FBl1"              "EPA(L)"            "GOR(R)"           
#>  [40] "AB(L)"             "aL(R)"             "PB(L3)"           
#>  [43] "NO(R)"             "LH(R)"             "IPS(R)"           
#>  [46] "AMMC"              "b'2(R)"            "PB(R1)"           
#>  [49] "PB(R3)"            "LX(L)"             "RUB(L)"           
#>  [52] "POC"               "EB"                "FBl5"             
#>  [55] "AME(R)"            "PB(R7)"            "a'3(R)"           
#>  [58] "a'2(R)"            "NO"                "gL(R)"            
#>  [61] "g1(R)"             "mALT(R)"           "NO3(R)"           
#>  [64] "PB(L7)"            "GNG"               "MB(R)"            
#>  [67] "a1(R)"             "PED(R)"            "g4(R)"            
#>  [70] "ROB(R)"            "PB(L9)"            "LO(R)"            
#>  [73] "ATL(R)"            "GC"                "a'L(L)"           
#>  [76] "RUB(R)"            "SIP(R)"            "LX(R)"            
#>  [79] "ME(R)"             "EBr3pw"            "g5(R)"            
#>  [82] "FBl2"              "FBl8"              "PB(R8)"           
#>  [85] "GOR(L)"            "AL(R)"             "EBr3am"           
#>  [88] "PB(R6)"            "LAL(-GA)(R)"       "FLA(R)"           
#>  [91] "VMNP"              "CRE(L)"            "bL(L)"            
#>  [94] "VES(R)"            "AB(R)"             "AOTU(R)"          
#>  [97] "SLP(R)"            "SNP(L)"            "SMP(R)"           
#> [100] "CX"                "SCL(R)"            "EBr5"             
#> [103] "PB"                "FBl3"              "VLNP(R)"          
#> [106] "GA(R)"             "gL(L)"             "CA(R)"            
#> [109] "NO2(R)"            "b'1(R)"            "ATL(L)"           
#> [112] "SPS(L)"            "SNP(R)"            "AVLP(R)"          
#> [115] "OL(R)"             "PB(R5)"            "vACA(R)"          
#> [118] "hemibrain"         "WED(R)"            "PB(L1)"           
#> [121] "AL(L)"             "FBl4"              "INP"              
#> [124] "NO(L)"             "PRW"               "CRE(-ROB,-RUB)(R)"
#> [127] "PB(R4)"            "SIP(L)"            "CRE(R)"           
#> [130] "PB(R9)"            "b'L(L)"            "a2(R)"            
#> [133] "LAL(R)"            "PLP(R)"            "BU(L)"            
#> [136] "b1(R)"             "PB(L4)"            "g3(R)"            
#> [139] "aL(L)"             "LAL(L)"            "AOT(R)"           
#> [142] "SMP(L)"            "lACA(R)"           "PB(L6)"           
#> [145] "bL(R)"             "SCL(L)"            "SPS(R)"           
#> [148] "b2(R)"             "dACA(R)"           "EBr3d"            
#> [151] "SAD"              
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "LH(R)"             "AL-VM7v(R)"        "PB"               
#>   [4] "AL-DP1l(R)"        "b'2(R)"            "FBl9"             
#>   [7] "AL-DL4(L)"         "gL(R)"             "LAL(R)"           
#>  [10] "AL-DM1(R)"         "PB(L2)"            "PB(L8)"           
#>  [13] "EBr3d"             "MB(+ACA)(R)"       "AL(R)"            
#>  [16] "SPS(R)"            "NO"                "AL-VM5v(R)"       
#>  [19] "PB(L3)"            "AL-VC1(R)"         "PB(L4)"           
#>  [22] "PB(R9)"            "AB(R)"             "LX(R)"            
#>  [25] "EBr2r4"            "AVLP(R)"           "SPS(L)"           
#>  [28] "AL-VM7v(L)"        "CAN(R)"            "MB(L)"            
#>  [31] "PB(R8)"            "SIP(R)"            "FBl6"             
#>  [34] "AL-DC1(L)"         "EPA(L)"            "PB(R2)"           
#>  [37] "AL-VM7d(L)"        "VES(R)"            "AL-VC2(R)"        
#>  [40] "CA(L)"             "PENP"              "AL-VA7m(R)"       
#>  [43] "b1(R)"             "AL-DL2d(R)"        "BU(L)"            
#>  [46] "dACA(R)"           "AL-DA4l(R)"        "AL-VP4(R)"        
#>  [49] "bL(L)"             "FBl2"              "a'L(L)"           
#>  [52] "AL-DL5(L)"         "EBr3pw"            "FBl8"             
#>  [55] "AB(L)"             "EBr3am"            "ATL(L)"           
#>  [58] "NO(R)"             "SNP(L)"            "SLP(R)"           
#>  [61] "AL-DM5(L)"         "AL-DM6(L)"         "g5(R)"            
#>  [64] "SIP(L)"            "SCL(R)"            "AL-DM5(R)"        
#>  [67] "AL-VC5(R)"         "LAL(L)"            "INP"              
#>  [70] "AL-DA1(R)"         "AL-DM3(R)"         "VLNP(R)"          
#>  [73] "mALT(R)"           "FBl1"              "AL-VA7l(R)"       
#>  [76] "lACA(R)"           "AL-DC3(R)"         "RUB(R)"           
#>  [79] "SMP(L)"            "OL(R)"             "PB(L6)"           
#>  [82] "AL-VP2(R)"         "SAD"               "FB"               
#>  [85] "bL(R)"             "AL-VA3(R)"         "AL-VA4(R)"        
#>  [88] "AL-VA6(R)"         "PB(R6)"            "b'L(R)"           
#>  [91] "AL-DL1(R)"         "AL-D(R)"           "AL-VM7d(R)"       
#>  [94] "ICL(L)"            "GNG"               "PB(L9)"           
#>  [97] "PVLP(R)"           "b'L(L)"            "AL-DA3(R)"        
#> [100] "CRE(-ROB,-RUB)(R)" "AMMC"              "PB(R4)"           
#> [103] "SCL(L)"            "mALT(L)"           "AL-VA6(L)"        
#> [106] "AL-VP3(R)"         "AL-VM4(R)"         "FB-column3"       
#> [109] "RUB(L)"            "PB(R7)"            "AL-VA2(R)"        
#> [112] "AL-DM1(L)"         "FBl7"              "PB(R3)"           
#> [115] "POC"               "a'3(R)"            "ME(R)"            
#> [118] "AL-DC4(L)"         "a'1(R)"            "FBl4"             
#> [121] "FBl5"              "AL-VM2(R)"         "AL-DA2(L)"        
#> [124] "AL-D(L)"           "CX"                "GOR(L)"           
#> [127] "AL-DP1m(R)"        "AL-DC1(R)"         "AL-VL1(R)"        
#> [130] "AL-VP1m(R)"        "hemibrain"         "AME(R)"           
#> [133] "AL-VL2p(R)"        "AL-VM1(R)"         "PB(L5)"           
#> [136] "EBr5"              "AL-DA3(L)"         "AL-VP1l(R)"       
#> [139] "IB"                "PLP(R)"            "GOR(R)"           
#> [142] "a2(R)"             "CRE(R)"            "IPS(R)"           
#> [145] "AL-V(R)"           "AL-DC2(L)"         "CRE(-RUB)(L)"     
#> [148] "AL-DL5(R)"         "AL-VM3(R)"         "AOT(R)"           
#> [151] "g1(R)"             "AL-DA2(R)"         "CA(R)"            
#> [154] "AL-DM3(L)"         "NO3(L)"            "a3(R)"            
#> [157] "g3(R)"             "LAL(-GA)(R)"       "ICL(R)"           
#> [160] "NO2(R)"            "aL(L)"             "MB(R)"            
#> [163] "AL-DC4(R)"         "AOTU(R)"           "VMNP"             
#> [166] "AL-DA4m(R)"        "AL-VP1d(R)"        "a1(R)"            
#> [169] "EBr1"              "FBl3"              "EBr6"             
#> [172] "NO1(R)"            "a'2(R)"            "PB(L7)"           
#> [175] "SMP(R)"            "GF(R)"             "PRW"              
#> [178] "b2(R)"             "VES(L)"            "vACA(R)"          
#> [181] "a'L(R)"            "ROB(R)"            "PB(R1)"           
#> [184] "AL-VP5(R)"         "PB(R5)"            "NO1(L)"           
#> [187] "BU(R)"             "NO(L)"             "g4(R)"            
#> [190] "EPA(R)"            "AL-VA1v(R)"        "b'1(R)"           
#> [193] "NO3(R)"            "AL-DM2(R)"         "LO(R)"            
#> [196] "AL-DM4(R)"         "AL-VA1d(R)"        "AL-VM5d(R)"       
#> [199] "aL(R)"             "EB"                "AL-DL2v(R)"       
#> [202] "AL-DL4(R)"         "GC"                "AL-DC2(R)"        
#> [205] "AL(L)"             "AL-DM2(L)"         "NO2(L)"           
#> [208] "PED(R)"            "WED(R)"            "ATL(R)"           
#> [211] "AL-DM6(R)"         "AL-VC3m(R)"        "AL-DP1m(L)"       
#> [214] "LX(L)"             "gL(L)"             "SAD(-AMMC)"       
#> [217] "GA(R)"             "FLA(R)"            "AL-VC3l(R)"       
#> [220] "CRE(L)"            "AL-DL3(R)"         "AL-VA5(R)"        
#> [223] "AL-VC4(R)"         "AL-VL2a(R)"        "PB(L1)"           
#> [226] "AL-DM4(L)"         "AL-DA4m(L)"        "LOP(R)"           
#> [229] "g2(R)"             "SNP(R)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-DA2(L)"         "AL-DM5(R)"         "EBr1"             
#>   [4] "LO(R)"             "AL-VA3(R)"         "CX"               
#>   [7] "EBr3pw"            "SLP(R)"            "AL-VM7v(R)"       
#>  [10] "vACA(R)"           "b'2(R)"            "a'L(R)"           
#>  [13] "AME(R)"            "MB(R)"             "AL-VM4(R)"        
#>  [16] "VLNP(R)"           "PB(L3)"            "PB(R4)"           
#>  [19] "AL-VL1(R)"         "SMP(L)"            "AL-DP1m(L)"       
#>  [22] "aL(L)"             "LAL(-GA)(R)"       "AL-DM6(L)"        
#>  [25] "CRE(-RUB)(L)"      "AL-VA1v(R)"        "a3(R)"            
#>  [28] "CA(L)"             "a2(R)"             "g2(R)"            
#>  [31] "a'2(R)"            "ICL(R)"            "INP"              
#>  [34] "AL-DC3(R)"         "AL-VP1d(R)"        "g3(R)"            
#>  [37] "b2(R)"             "PB(L8)"            "FBl5"             
#>  [40] "AL-DL2v(R)"        "AL-DM3(R)"         "LAL(L)"           
#>  [43] "PB(L7)"            "AL-DA2(R)"         "EBr2r4"           
#>  [46] "EBr5"              "POC"               "AL-DL3(R)"        
#>  [49] "AL-VA6(R)"         "PRW"               "AL-VP1m(R)"       
#>  [52] "gL(L)"             "AL(R)"             "AL-DL4(R)"        
#>  [55] "AL-DP1m(R)"        "AL-V(R)"           "a'3(R)"           
#>  [58] "GC"                "AL-VA2(R)"         "EBr3am"           
#>  [61] "PB(R6)"            "AL-DC1(R)"         "AL-VA5(R)"        
#>  [64] "a1(R)"             "GOR(L)"            "mALT(R)"          
#>  [67] "NO3(R)"            "SNP(R)"            "AL-DL4(L)"        
#>  [70] "AL-DL1(R)"         "AL-VC1(R)"         "AL-DA4m(L)"       
#>  [73] "AL-VL2p(R)"        "SAD(-AMMC)"        "SPS(R)"           
#>  [76] "NO(L)"             "AL-VA7m(R)"        "IB"               
#>  [79] "SIP(L)"            "PB(R5)"            "AL-DA3(L)"        
#>  [82] "AL-VM7v(L)"        "AL-DA4m(R)"        "AL-DM2(R)"        
#>  [85] "AL-VA4(R)"         "AL-DC4(L)"         "AL-VM1(R)"        
#>  [88] "FBl3"              "AL-D(R)"           "AL-VP3(R)"        
#>  [91] "AL-VP2(R)"         "PB(L5)"            "SAD"              
#>  [94] "NO1(R)"            "PLP(R)"            "AL-VA6(L)"        
#>  [97] "CRE(-ROB,-RUB)(R)" "AMMC"              "AL-DM2(L)"        
#> [100] "g5(R)"             "VES(R)"            "FBl2"             
#> [103] "AL-DL5(R)"         "PVLP(R)"           "IPS(R)"           
#> [106] "OL(R)"             "AL-DM1(R)"         "AL-DM4(R)"        
#> [109] "AL-VL2a(R)"        "AL-VP5(R)"         "dACA(R)"          
#> [112] "g4(R)"             "CA(R)"             "AL-DL5(L)"        
#> [115] "CRE(L)"            "AL-VP4(R)"         "ATL(L)"           
#> [118] "CRE(R)"            "EPA(R)"            "a'1(R)"           
#> [121] "AL-DL2d(R)"        "AL-VM5d(R)"        "WED(R)"           
#> [124] "LX(R)"             "PB(L6)"            "gL(R)"            
#> [127] "PB(R1)"            "AVLP(R)"           "AL-VM3(R)"        
#> [130] "BU(R)"             "AL-D(L)"           "SNP(L)"           
#> [133] "AL-DP1l(R)"        "RUB(R)"            "ICL(L)"           
#> [136] "MB(+ACA)(R)"       "FBl7"              "ATL(R)"           
#> [139] "EPA(L)"            "AL-VC3l(R)"        "mALT(L)"          
#> [142] "GA(R)"             "FB"                "FBl4"             
#> [145] "aL(R)"             "b'L(R)"            "AL(L)"            
#> [148] "AL-VM7d(L)"        "AL-DA1(R)"         "AL-VA7l(R)"       
#> [151] "AL-VM7d(R)"        "NO"                "FBl8"             
#> [154] "hemibrain"         "MB(L)"             "PB(L9)"           
#> [157] "PB(R2)"            "b'1(R)"            "BU(L)"            
#> [160] "PB(L1)"            "PB(R8)"            "VMNP"             
#> [163] "PB(R7)"            "g1(R)"             "LAL(R)"           
#> [166] "AL-DA4l(R)"        "AL-VM5v(R)"        "AL-DC2(L)"        
#> [169] "GNG"               "AL-VP1l(R)"        "AL-DM4(L)"        
#> [172] "PB(R3)"            "AL-VC5(R)"         "PB(L2)"           
#> [175] "PENP"              "lACA(R)"           "PED(R)"           
#> [178] "SCL(R)"            "NO(R)"             "FBl9"             
#> [181] "AOT(R)"            "NO3(L)"            "AL-DA3(R)"        
#> [184] "AB(L)"             "SCL(L)"            "ME(R)"            
#> [187] "NO2(R)"            "AL-DM1(L)"         "SIP(R)"           
#> [190] "AL-DC2(R)"         "AL-VC3m(R)"        "AL-VC4(R)"        
#> [193] "SMP(R)"            "GF(R)"             "SPS(L)"           
#> [196] "VES(L)"            "AL-DM3(L)"         "ROB(R)"           
#> [199] "LX(L)"             "AL-DM6(R)"         "LH(R)"            
#> [202] "FBl1"              "b1(R)"             "AL-VA1d(R)"       
#> [205] "NO2(L)"            "NO1(L)"            "PB"               
#> [208] "b'L(L)"            "FLA(R)"            "AL-DM5(L)"        
#> [211] "bL(L)"             "FBl6"              "EBr3d"            
#> [214] "FB-column3"        "bL(R)"             "AL-DC1(L)"        
#> [217] "a'L(L)"            "AL-DC4(R)"         "EBr6"             
#> [220] "PB(R9)"            "RUB(L)"            "GOR(R)"           
#> [223] "AL-VC2(R)"         "EB"                "AL-VM2(R)"        
#> [226] "LOP(R)"            "AOTU(R)"           "AB(R)"            
#> [229] "CAN(R)"            "PB(L4)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method   path                                     
#> [1,] "HEAD"   "/api/*"                                 
#> [2,] "REPORT" "/api/*"                                 
#> [3,] "GET"    "/api/v:ver/dbmeta/datasets"             
#> [4,] "GET"    "/api/v:ver/npexplorer/simpleconnections"
#> [5,] "GET"    "/login"                                 
#> [6,] "GET"    "/api/v:ver/npexplorer/distribution"     
#>      name                                                                
#> [1,] "github.com/labstack/echo.glob..func1"                              
#> [2,] "github.com/labstack/echo.glob..func1"                              
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [5,] "github.com/janelia-flyem/echo-secure.loginHandler"                 
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }