Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "g2(R)"             "SPS(L)"            "AL-VA6(L)"        
#>   [4] "AL-VA1d(R)"        "FBl8"              "bL(R)"            
#>   [7] "AL-DM4(R)"         "AMMC"              "FLA(R)"           
#>  [10] "AL-DM6(L)"         "EBr3pw"            "AL-DC1(R)"        
#>  [13] "INP"               "ATL(L)"            "PRW"              
#>  [16] "FB"                "EB"                "AL-DP1m(L)"       
#>  [19] "PENP"              "EPA(R)"            "AL-VP1m(R)"       
#>  [22] "mALT(L)"           "SMP(R)"            "IPS(R)"           
#>  [25] "AL-VA1v(R)"        "AL-VA2(R)"         "AL-VA7l(R)"       
#>  [28] "SIP(L)"            "BU(L)"             "LAL(R)"           
#>  [31] "AL-VM1(R)"         "AOT(R)"            "AL-VL2p(R)"       
#>  [34] "b'2(R)"            "PB(L3)"            "aL(R)"            
#>  [37] "CA(R)"             "NO1(R)"            "AL-DL3(R)"        
#>  [40] "AL-DM6(R)"         "AL-VM5d(R)"        "SNP(L)"           
#>  [43] "AL-VM7d(R)"        "b2(R)"             "MB(R)"            
#>  [46] "bL(L)"             "PB(L2)"            "AL-VP4(R)"        
#>  [49] "PB(L1)"            "AL-DM5(L)"         "FBl5"             
#>  [52] "AL-VL2a(R)"        "AL-DA2(R)"         "AL-DL4(L)"        
#>  [55] "NO"                "LO(R)"             "OL(R)"            
#>  [58] "ICL(L)"            "VES(L)"            "AL-DL5(R)"        
#>  [61] "AL-VM2(R)"         "PB(R1)"            "PB(R2)"           
#>  [64] "b'1(R)"            "CRE(-RUB)(L)"      "AL-VP1l(R)"       
#>  [67] "AL-VP1d(R)"        "AB(L)"             "GF(R)"            
#>  [70] "EBr1"              "AL-DP1m(R)"        "NO2(L)"           
#>  [73] "vACA(R)"           "EBr5"              "NO3(L)"           
#>  [76] "AL-VC3m(R)"        "AL-VM5v(R)"        "BU(R)"            
#>  [79] "CRE(L)"            "a3(R)"             "PB(L6)"           
#>  [82] "AL-VA6(R)"         "AL-VL1(R)"         "a'3(R)"           
#>  [85] "g4(R)"             "AL-DM1(L)"         "AL-DC2(R)"        
#>  [88] "AL-DL2d(R)"        "PB(R8)"            "AL-DM4(L)"        
#>  [91] "FBl9"              "AL-DP1l(R)"        "AL-DL1(R)"        
#>  [94] "GA(R)"             "SAD"               "a'L(R)"           
#>  [97] "VES(R)"            "PB(R7)"            "AL-DM2(L)"        
#> [100] "PVLP(R)"           "AL-VM7v(L)"        "NO(R)"            
#> [103] "CA(L)"             "a'2(R)"            "AL-DC1(L)"        
#> [106] "AL(R)"             "AL-VC1(R)"         "AL-DM1(R)"        
#> [109] "AL-DM3(R)"         "POC"               "NO3(R)"           
#> [112] "ICL(R)"            "GNG"               "b'L(L)"           
#> [115] "AL-VM7d(L)"        "b1(R)"             "AB(R)"            
#> [118] "PB(R4)"            "AL-VP2(R)"         "AL-DA3(L)"        
#> [121] "AL-VP5(R)"         "SCL(L)"            "a2(R)"            
#> [124] "SIP(R)"            "PB"                "EBr2r4"           
#> [127] "g5(R)"             "AL-VC4(R)"         "GC"               
#> [130] "EBr3d"             "AL(L)"             "AL-DA3(R)"        
#> [133] "AL-VM7v(R)"        "AL-VP3(R)"         "VLNP(R)"          
#> [136] "PB(L5)"            "SCL(R)"            "FBl1"             
#> [139] "NO2(R)"            "aL(L)"             "CX"               
#> [142] "LAL(L)"            "PED(R)"            "AL-DM5(R)"        
#> [145] "VMNP"              "gL(L)"             "AL-DA2(L)"        
#> [148] "LH(R)"             "a'1(R)"            "a'L(L)"           
#> [151] "AL-D(R)"           "SMP(L)"            "SPS(R)"           
#> [154] "AL-DL5(L)"         "ROB(R)"            "dACA(R)"          
#> [157] "AL-DC2(L)"         "AL-VM3(R)"         "LOP(R)"           
#> [160] "AL-DM3(L)"         "MB(+ACA)(R)"       "AL-V(R)"          
#> [163] "mALT(R)"           "AL-DL2v(R)"        "hemibrain"        
#> [166] "g1(R)"             "IB"                "FBl2"             
#> [169] "EPA(L)"            "AL-VA4(R)"         "PB(R6)"           
#> [172] "GOR(L)"            "EBr6"              "PB(L9)"           
#> [175] "lACA(R)"           "AL-D(L)"           "PB(L7)"           
#> [178] "CAN(R)"            "AL-DA1(R)"         "AL-VC5(R)"        
#> [181] "AL-DC4(L)"         "IB112"             "FBl7"             
#> [184] "AL-VC3l(R)"        "AL-DC3(R)"         "g3(R)"            
#> [187] "PB(L8)"            "PB(R3)"            "CRE(R)"           
#> [190] "FBl6"              "LX(L)"             "AL-VA3(R)"        
#> [193] "b'L(R)"            "AME(R)"            "NO(L)"            
#> [196] "SLP(R)"            "RUB(R)"            "CRE(-ROB,-RUB)(R)"
#> [199] "WED(R)"            "gL(R)"             "RUB(L)"           
#> [202] "PB(R5)"            "ATL(R)"            "AL-VC2(R)"        
#> [205] "AL-VM4(R)"         "MB(L)"             "SNP(R)"           
#> [208] "AL-DA4m(L)"        "PB(R9)"            "SAD(-AMMC)"       
#> [211] "AL-DM2(R)"         "AL-DL4(R)"         "EBr3am"           
#> [214] "LAL(-GA)(R)"       "FBl4"              "AL-DA4l(R)"       
#> [217] "AL-DA4m(R)"        "AOTU(R)"           "ME(R)"            
#> [220] "AL-VA5(R)"         "NO1(L)"            "AL-DC4(R)"        
#> [223] "AL-VA7m(R)"        "AVLP(R)"           "PB(L4)"           
#> [226] "a1(R)"             "PLP(R)"            "LX(R)"            
#> [229] "FB-column3"        "FBl3"              "GOR(R)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "g1(R)"             "EBr6"              "g5(R)"            
#>   [4] "AME(R)"            "PB(L2)"            "PB(L3)"           
#>   [7] "a'L(R)"            "a'1(R)"            "NO3(L)"           
#>  [10] "FBl3"              "PB(R5)"            "FBl9"             
#>  [13] "GNG"               "PB(L4)"            "SCL(L)"           
#>  [16] "SMP(L)"            "g3(R)"             "LX(R)"            
#>  [19] "ME(R)"             "AVLP(R)"           "ICL(R)"           
#>  [22] "a1(R)"             "b2(R)"             "PB(R8)"           
#>  [25] "ROB(R)"            "VES(R)"            "RUB(R)"           
#>  [28] "vACA(R)"           "GC"                "PB(L6)"           
#>  [31] "PB(L9)"            "FBl7"              "RUB(L)"           
#>  [34] "PB(L5)"            "FB"                "LX(L)"            
#>  [37] "LOP(R)"            "a'3(R)"            "PED(R)"           
#>  [40] "dACA(R)"           "CRE(L)"            "LO(R)"            
#>  [43] "FBl4"              "INP"               "ATL(L)"           
#>  [46] "CA(L)"             "NO2(L)"            "AOTU(R)"          
#>  [49] "MB(L)"             "ICL(L)"            "FBl8"             
#>  [52] "CAN(R)"            "mALT(L)"           "hemibrain"        
#>  [55] "GOR(L)"            "a'2(R)"            "NO(R)"            
#>  [58] "PB(L7)"            "PB(R3)"            "aL(R)"            
#>  [61] "EBr3pw"            "b'2(R)"            "SIP(R)"           
#>  [64] "bL(L)"             "ATL(R)"            "SMP(R)"           
#>  [67] "SNP(R)"            "PB(R2)"            "CRE(R)"           
#>  [70] "g4(R)"             "LAL(R)"            "b'1(R)"           
#>  [73] "SAD"               "mALT(R)"           "BU(L)"            
#>  [76] "PB(R9)"            "b'L(L)"            "OL(R)"            
#>  [79] "CX"                "SPS(R)"            "gL(R)"            
#>  [82] "FBl6"              "NO2(R)"            "SLP(R)"           
#>  [85] "AOT(R)"            "AB(R)"             "LAL(L)"           
#>  [88] "SIP(L)"            "NO1(L)"            "g2(R)"            
#>  [91] "PLP(R)"            "EB"                "LAL(-GA)(R)"      
#>  [94] "SPS(L)"            "NO(L)"             "GA(R)"            
#>  [97] "PB"                "AL(R)"             "SCL(R)"           
#> [100] "SAD(-AMMC)"        "CA(R)"             "GF(R)"            
#> [103] "a2(R)"             "PB(L8)"            "VMNP"             
#> [106] "FBl5"              "VES(L)"            "PVLP(R)"          
#> [109] "PB(R7)"            "NO3(R)"            "EPA(L)"           
#> [112] "BU(R)"             "EPA(R)"            "PB(R1)"           
#> [115] "b1(R)"             "GOR(R)"            "a3(R)"            
#> [118] "WED(R)"            "MB(+ACA)(R)"       "FBl2"             
#> [121] "AMMC"              "FBl1"              "PB(R4)"           
#> [124] "bL(R)"             "EBr3am"            "IB"               
#> [127] "PB(R6)"            "EBr1"              "FLA(R)"           
#> [130] "AL(L)"             "SNP(L)"            "LH(R)"            
#> [133] "lACA(R)"           "NO1(R)"            "EBr2r4"           
#> [136] "gL(L)"             "IPS(R)"            "MB(R)"            
#> [139] "AB(L)"             "CRE(-ROB,-RUB)(R)" "PENP"             
#> [142] "b'L(R)"            "EBr5"              "a'L(L)"           
#> [145] "VLNP(R)"           "POC"               "NO"               
#> [148] "PRW"               "PB(L1)"            "EBr3d"            
#> [151] "aL(L)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "FBl7"              "mALT(R)"           "SNP(L)"           
#>   [4] "AL-DA3(L)"         "FBl2"              "PB(L1)"           
#>   [7] "LX(R)"             "MB(+ACA)(R)"       "PB(R3)"           
#>  [10] "AL-DM6(R)"         "PB(L8)"            "ROB(R)"           
#>  [13] "aL(R)"             "MB(R)"             "CAN(R)"           
#>  [16] "GOR(R)"            "PB(R7)"            "LX(L)"            
#>  [19] "EPA(L)"            "NO(L)"             "AL-VC5(R)"        
#>  [22] "IB"                "PB(L5)"            "AL-DM2(L)"        
#>  [25] "gL(R)"             "aL(L)"             "SMP(R)"           
#>  [28] "PB"                "VMNP"              "AL-DM4(R)"        
#>  [31] "AL-VP2(R)"         "VLNP(R)"           "GOR(L)"           
#>  [34] "PB(L4)"            "AL-DC1(R)"         "AL-VM2(R)"        
#>  [37] "EBr5"              "PB(L9)"            "AL-DC4(R)"        
#>  [40] "AL-DL4(R)"         "AL-VC3m(R)"        "SAD"              
#>  [43] "WED(R)"            "FBl4"              "AL-DL2v(R)"       
#>  [46] "PED(R)"            "NO1(L)"            "SAD(-AMMC)"       
#>  [49] "CA(R)"             "VES(L)"            "b'1(R)"           
#>  [52] "a'L(L)"            "AL-VL2p(R)"        "PB(L2)"           
#>  [55] "AL-DM6(L)"         "a'1(R)"            "EBr3am"           
#>  [58] "ME(R)"             "PB(R5)"            "dACA(R)"          
#>  [61] "AL-D(R)"           "AL-VA3(R)"         "RUB(R)"           
#>  [64] "g1(R)"             "AL-DL4(L)"         "AL-DP1l(R)"       
#>  [67] "AL-VM5v(R)"        "POC"               "PB(R8)"           
#>  [70] "AL-DM1(R)"         "FBl6"              "AL-VP1m(R)"       
#>  [73] "LOP(R)"            "b2(R)"             "bL(L)"            
#>  [76] "LH(R)"             "AL-DL3(R)"         "AL-DP1m(R)"       
#>  [79] "vACA(R)"           "a1(R)"             "b'L(R)"           
#>  [82] "AL-DC2(L)"         "AL-VM1(R)"         "AL-VM7d(R)"       
#>  [85] "AL-VP1l(R)"        "SCL(L)"            "a'2(R)"           
#>  [88] "AL-DL1(R)"         "ICL(R)"            "AL-VM3(R)"        
#>  [91] "NO2(L)"            "CRE(R)"            "AL-DC1(L)"        
#>  [94] "mALT(L)"           "NO3(R)"            "PB(R9)"           
#>  [97] "FBl3"              "AL-DL2d(R)"        "SMP(L)"           
#> [100] "VES(R)"            "PENP"              "AL-VA5(R)"        
#> [103] "AL-VC4(R)"         "AL-VP5(R)"         "AL-D(L)"          
#> [106] "SLP(R)"            "AL-DA4m(R)"        "BU(L)"            
#> [109] "EBr3d"             "SNP(R)"            "PB(L7)"           
#> [112] "hemibrain"         "EBr6"              "AL-DC4(L)"        
#> [115] "EBr1"              "g4(R)"             "AL-DC2(R)"        
#> [118] "AL-VA4(R)"         "b'2(R)"            "gL(L)"            
#> [121] "CRE(L)"            "EBr3pw"            "GF(R)"            
#> [124] "AL-VM7v(L)"        "AL-DM5(L)"         "AL-DM2(R)"        
#> [127] "lACA(R)"           "AL-DM3(L)"         "AVLP(R)"          
#> [130] "AB(L)"             "AL-DM4(L)"         "AL-DP1m(L)"       
#> [133] "AL-DL5(R)"         "AL-VP1d(R)"        "MB(L)"            
#> [136] "NO3(L)"            "SIP(L)"            "AL-VM7v(R)"       
#> [139] "CA(L)"             "PB(R6)"            "AL-VC1(R)"        
#> [142] "EPA(R)"            "NO(R)"             "b1(R)"            
#> [145] "LAL(R)"            "AL-VA2(R)"         "AL-VA7m(R)"       
#> [148] "g3(R)"             "PB(L3)"            "ATL(L)"           
#> [151] "AL-DA1(R)"         "a2(R)"             "FB-column3"       
#> [154] "FBl1"              "GC"                "SPS(L)"           
#> [157] "PB(R1)"            "ICL(L)"            "CRE(-ROB,-RUB)(R)"
#> [160] "LAL(L)"            "g2(R)"             "AL-DA4l(R)"       
#> [163] "AL-VL1(R)"         "AOT(R)"            "PB(R2)"           
#> [166] "AME(R)"            "OL(R)"             "NO"               
#> [169] "PB(L6)"            "AL-VL2a(R)"        "AB(R)"            
#> [172] "AL-DA3(R)"         "AL-DM5(R)"         "g5(R)"            
#> [175] "FBl5"              "AL-VA7l(R)"        "AMMC"             
#> [178] "PLP(R)"            "a'3(R)"            "AL(L)"            
#> [181] "AL-DL5(L)"         "a3(R)"             "AL-DM3(R)"        
#> [184] "AL-VC3l(R)"        "GA(R)"             "FLA(R)"           
#> [187] "EB"                "AL-DA2(L)"         "AL-VA6(L)"        
#> [190] "FB"                "NO1(R)"            "PB(R4)"           
#> [193] "AL-VA1d(R)"        "AL-VM5d(R)"        "AL-V(R)"          
#> [196] "AL-VP4(R)"         "SCL(R)"            "FBl8"             
#> [199] "AL-DM1(L)"         "LO(R)"             "AL-DC3(R)"        
#> [202] "AL-VA6(R)"         "AOTU(R)"           "SPS(R)"           
#> [205] "bL(R)"             "PRW"               "SIP(R)"           
#> [208] "b'L(L)"            "CX"                "NO2(R)"           
#> [211] "AL-DA2(R)"         "AL-VA1v(R)"        "AL-VC2(R)"        
#> [214] "LAL(-GA)(R)"       "BU(R)"             "PVLP(R)"          
#> [217] "RUB(L)"            "AL(R)"             "AL-VP3(R)"        
#> [220] "FBl9"              "IPS(R)"            "GNG"              
#> [223] "a'L(R)"            "AL-VM7d(L)"        "AL-DA4m(L)"       
#> [226] "INP"               "EBr2r4"            "ATL(R)"           
#> [229] "CRE(-RUB)(L)"      "AL-VM4(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL(L)"             "AL-DA2(L)"         "GOR(L)"           
#>   [4] "NO"                "INP"               "bL(R)"            
#>   [7] "AL-VL2a(R)"        "LAL(-GA)(R)"       "VMNP"             
#>  [10] "AL-VA6(L)"         "GOR(R)"            "AL-VC3l(R)"       
#>  [13] "PB(R8)"            "PB(R2)"            "PB(L4)"           
#>  [16] "AMMC"              "BU(L)"             "AL-VM7v(R)"       
#>  [19] "EBr5"              "AL-D(L)"           "AL-VP5(R)"        
#>  [22] "dACA(R)"           "NO1(R)"            "SCL(L)"           
#>  [25] "mALT(R)"           "AL-VM7v(L)"        "AL-VP1m(R)"       
#>  [28] "PLP(R)"            "AL-DM3(L)"         "AL-DP1m(L)"       
#>  [31] "FBl8"              "FBl9"              "a3(R)"            
#>  [34] "AL-VM4(R)"         "AL-VM5v(R)"        "lACA(R)"          
#>  [37] "EBr3pw"            "IPS(R)"            "AL-DA4l(R)"       
#>  [40] "AL-DC2(R)"         "AL-DM5(R)"         "SNP(R)"           
#>  [43] "AL-DM6(L)"         "NO2(R)"            "AL-VA2(R)"        
#>  [46] "LO(R)"             "FBl7"              "PB(R3)"           
#>  [49] "AL-DM2(R)"         "AL-VP1l(R)"        "SMP(R)"           
#>  [52] "AL-DL4(L)"         "AL-DL5(L)"         "a'1(R)"           
#>  [55] "AL-DA2(R)"         "AL-VC2(R)"         "ME(R)"            
#>  [58] "AL-DM1(L)"         "AL-VA1v(R)"        "AL-VP3(R)"        
#>  [61] "WED(R)"            "g1(R)"             "ICL(L)"           
#>  [64] "PB(R1)"            "CA(R)"             "LH(R)"            
#>  [67] "AL-DA3(R)"         "AL-VP1d(R)"        "GA(R)"            
#>  [70] "FBl5"              "PB(R4)"            "VES(L)"           
#>  [73] "RUB(R)"            "AL-VM1(R)"         "b'2(R)"           
#>  [76] "b2(R)"             "AL-DC1(L)"         "AB(R)"            
#>  [79] "b'1(R)"            "AL-VA5(R)"         "AL-VC3m(R)"       
#>  [82] "PB(L1)"            "AL-DC2(L)"         "CRE(-RUB)(L)"     
#>  [85] "FBl1"              "ATL(R)"            "b'L(L)"           
#>  [88] "EBr3am"            "SIP(R)"            "NO(R)"            
#>  [91] "AVLP(R)"           "AL(R)"             "AL-VA1d(R)"       
#>  [94] "AL-VA3(R)"         "AL-VM3(R)"         "FB"               
#>  [97] "FLA(R)"            "EPA(L)"            "LX(R)"            
#> [100] "EB"                "PB(R9)"            "PB(L2)"           
#> [103] "AL-DC4(L)"         "ROB(R)"            "FBl3"             
#> [106] "AOTU(R)"           "PED(R)"            "EBr6"             
#> [109] "AL-DA4m(R)"        "FBl2"              "AL-VA6(R)"        
#> [112] "a1(R)"             "SIP(L)"            "g2(R)"            
#> [115] "PB(L8)"            "PB(L9)"            "OL(R)"            
#> [118] "SMP(L)"            "SPS(L)"            "aL(L)"            
#> [121] "EPA(R)"            "AL-DA1(R)"         "AL-DP1l(R)"       
#> [124] "gL(R)"             "SPS(R)"            "PB(L7)"           
#> [127] "AL-DM4(R)"         "ATL(L)"            "PENP"             
#> [130] "hemibrain"         "PB"                "FB-column3"       
#> [133] "LX(L)"             "AL-VL1(R)"         "PRW"              
#> [136] "AL-DL5(R)"         "IB"                "PB(R6)"           
#> [139] "AL-DM1(R)"         "CX"                "a2(R)"            
#> [142] "AL-VM2(R)"         "vACA(R)"           "PB(L5)"           
#> [145] "NO2(L)"            "GNG"               "a'L(L)"           
#> [148] "VLNP(R)"           "g3(R)"             "AME(R)"           
#> [151] "POC"               "bL(L)"             "g5(R)"            
#> [154] "VES(R)"            "AB(L)"             "AL-VM7d(L)"       
#> [157] "FBl6"              "b'L(R)"            "AL-DL4(R)"        
#> [160] "AL-DP1m(R)"        "AL-VC4(R)"         "PB(R5)"           
#> [163] "AL-DM5(L)"         "FBl4"              "AL-DL3(R)"        
#> [166] "gL(L)"             "SAD(-AMMC)"        "AL-DL2v(R)"       
#> [169] "AL-DM6(R)"         "AL-VC1(R)"         "NO3(R)"           
#> [172] "g4(R)"             "AL-DA3(L)"         "NO3(L)"           
#> [175] "SNP(L)"            "PB(R7)"            "MB(R)"            
#> [178] "AL-VC5(R)"         "AL-DL1(R)"         "AL-DL2d(R)"       
#> [181] "aL(R)"             "EBr3d"             "b1(R)"            
#> [184] "AL-DC3(R)"         "GF(R)"             "MB(L)"            
#> [187] "AL-VM7d(R)"        "a'2(R)"            "SAD"              
#> [190] "GC"                "AL-DM2(L)"         "AOT(R)"           
#> [193] "NO1(L)"            "CRE(R)"            "SCL(R)"           
#> [196] "BU(R)"             "CAN(R)"            "EBr2r4"           
#> [199] "a'L(R)"            "LOP(R)"            "AL-V(R)"          
#> [202] "AL-VP4(R)"         "AL-VP2(R)"         "PVLP(R)"          
#> [205] "LAL(R)"            "AL-VM5d(R)"        "MB(+ACA)(R)"      
#> [208] "ICL(R)"            "NO(L)"             "SLP(R)"           
#> [211] "AL-DC1(R)"         "CRE(-ROB,-RUB)(R)" "LAL(L)"           
#> [214] "EBr1"              "AL-DM3(R)"         "a'3(R)"           
#> [217] "mALT(L)"           "CA(L)"             "AL-VA7l(R)"       
#> [220] "AL-VA7m(R)"        "PB(L3)"            "PB(L6)"           
#> [223] "CRE(L)"            "RUB(L)"            "AL-D(R)"          
#> [226] "AL-VA4(R)"         "AL-DA4m(L)"        "AL-DM4(L)"        
#> [229] "AL-DC4(R)"         "AL-VL2p(R)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                
#> [1,] "GET"  "/api/v:ver/npexplorer/distribution"
#> [2,] "GET"  "/api/npexplorer/completeness"      
#> [3,] "GET"  "/api/roimeshes/mesh/:dataset/:roi" 
#> [4,] "GET"  "/api"                              
#> [5,] "PUT"  "/api"                              
#> [6,] "POST" "/api/v:ver/custom/custom"          
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCompleteness-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/roimeshes.masterAPI.getMesh-fm"            
#> [4,] "github.com/labstack/echo/v4.glob..func1"                                                   
#> [5,] "github.com/labstack/echo/v4.glob..func1"                                                   
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }