Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DC2(R)"         "AL-DL2d(R)"        "AL-VA7l(R)"       
#>   [4] "FLA(R)"            "PB(L4)"            "NO2(L)"           
#>   [7] "AOTU(R)"           "EBr6"              "SPS(L)"           
#>  [10] "VES(L)"            "IB"                "PENP"             
#>  [13] "AL-D(L)"           "AL-VL2a(R)"        "POC"              
#>  [16] "AL-VL2p(R)"        "CRE(-ROB,-RUB)(R)" "PRW"              
#>  [19] "AL-DC1(L)"         "CRE(L)"            "bL(L)"            
#>  [22] "ICL(L)"            "b'L(R)"            "AL-DC4(L)"        
#>  [25] "AL-DP1m(L)"        "AL-D(R)"           "b'2(R)"           
#>  [28] "a1(R)"             "PB(R8)"            "AL-DA2(R)"        
#>  [31] "AL-DC3(R)"         "AL-VP1d(R)"        "AL-DL2v(R)"       
#>  [34] "AL-DL4(R)"         "AL-DM2(R)"         "AL-DP1l(R)"       
#>  [37] "AL-VC2(R)"         "PB(L8)"            "AL-DL4(L)"        
#>  [40] "g3(R)"             "g4(R)"             "g5(R)"            
#>  [43] "RUB(R)"            "b'1(R)"            "AL-DC2(L)"        
#>  [46] "AL-DM4(L)"         "AVLP(R)"           "AL-DM3(R)"        
#>  [49] "aL(R)"             "NO(R)"             "lACA(R)"          
#>  [52] "MB(+ACA)(R)"       "ICL(R)"            "AL-DM1(R)"        
#>  [55] "AL-DM1(L)"         "AL-VA4(R)"         "AL-VC1(R)"        
#>  [58] "VLNP(R)"           "GC"                "AL-DM6(R)"        
#>  [61] "AL-VC4(R)"         "mALT(R)"           "AL-DM6(L)"        
#>  [64] "SAD(-AMMC)"        "a'L(L)"            "LO(R)"            
#>  [67] "AL-DP1m(R)"        "AL-VM7v(L)"        "AL-DA1(R)"        
#>  [70] "AL-DM5(R)"         "PB(R5)"            "AL-DM4(R)"        
#>  [73] "AL-VM7v(R)"        "EBr2r4"            "AL-VM5d(R)"       
#>  [76] "b2(R)"             "EPA(R)"            "AL-VP4(R)"        
#>  [79] "SMP(R)"            "a2(R)"             "PB(L7)"           
#>  [82] "PB(R3)"            "SLP(R)"            "AL-VM1(R)"        
#>  [85] "AL-VP2(R)"         "AL-DA4m(L)"        "AL-DL5(L)"        
#>  [88] "NO3(L)"            "PB(R1)"            "AL-DA4m(R)"       
#>  [91] "AOT(R)"            "AMMC"              "NO1(L)"           
#>  [94] "EBr3d"             "GOR(R)"            "FBl8"             
#>  [97] "AL(R)"             "AL-VP1m(R)"        "AL-VM4(R)"        
#> [100] "AL(L)"             "PB(R9)"            "b1(R)"            
#> [103] "AME(R)"            "AL-DA4l(R)"        "dACA(R)"          
#> [106] "AL-VM7d(L)"        "OL(R)"             "bL(R)"            
#> [109] "AL-VP5(R)"         "SMP(L)"            "LX(R)"            
#> [112] "FB-column3"        "aL(L)"             "PB(R4)"           
#> [115] "a'L(R)"            "FBl4"              "NO2(R)"           
#> [118] "AL-VA1d(R)"        "AL-VP3(R)"         "FB"               
#> [121] "AL-DC4(R)"         "AL-VM3(R)"         "a'3(R)"           
#> [124] "CRE(-RUB)(L)"      "GNG"               "b'L(L)"           
#> [127] "AL-DL3(R)"         "SAD"               "MB(R)"            
#> [130] "AL-VC3l(R)"        "FBl9"              "INP"              
#> [133] "FBl1"              "FBl6"              "AL-DA2(L)"        
#> [136] "FBl2"              "CA(R)"             "PB(R6)"           
#> [139] "PB(L3)"            "GA(R)"             "WED(R)"           
#> [142] "BU(R)"             "PB(L6)"            "EPA(L)"           
#> [145] "AL-VC5(R)"         "LOP(R)"            "AL-VA6(L)"        
#> [148] "FBl5"              "AL-VM2(R)"         "AL-DC1(R)"        
#> [151] "AL-VA6(R)"         "AL-VA7m(R)"        "AL-VP1l(R)"       
#> [154] "mALT(L)"           "ME(R)"             "EB"               
#> [157] "NO"                "NO(L)"             "AL-DA3(R)"        
#> [160] "BU(L)"             "PB(L9)"            "ROB(R)"           
#> [163] "EBr3pw"            "AL-VM7d(R)"        "PB(L2)"           
#> [166] "EBr3am"            "SIP(L)"            "AL-DM2(L)"        
#> [169] "VES(R)"            "CAN(R)"            "RUB(L)"           
#> [172] "EBr5"              "PB(L1)"            "a'2(R)"           
#> [175] "gL(L)"             "FBl7"              "AL-VC3m(R)"       
#> [178] "SNP(R)"            "a'1(R)"            "IB112"            
#> [181] "FBl3"              "LX(L)"             "AL-DL5(R)"        
#> [184] "AL-VA5(R)"         "AL-VL1(R)"         "GOR(L)"           
#> [187] "PB"                "ATL(R)"            "a3(R)"            
#> [190] "AL-VA2(R)"         "PB(R7)"            "NO3(R)"           
#> [193] "SIP(R)"            "PB(L5)"            "CX"               
#> [196] "AB(R)"             "SPS(R)"            "CRE(R)"           
#> [199] "MB(L)"             "gL(R)"             "IPS(R)"           
#> [202] "LAL(-GA)(R)"       "SCL(R)"            "EBr1"             
#> [205] "PVLP(R)"           "LH(R)"             "hemibrain"        
#> [208] "AL-DL1(R)"         "g1(R)"             "AL-DM3(L)"        
#> [211] "SNP(L)"            "AL-V(R)"           "vACA(R)"          
#> [214] "GF(R)"             "LAL(R)"            "AL-VA3(R)"        
#> [217] "NO1(R)"            "PLP(R)"            "VMNP"             
#> [220] "PB(R2)"            "AL-VA1v(R)"        "CA(L)"            
#> [223] "AL-DM5(L)"         "AB(L)"             "SCL(L)"           
#> [226] "ATL(L)"            "LAL(L)"            "PED(R)"           
#> [229] "g2(R)"             "AL-DA3(L)"         "AL-VM5v(R)"       
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "NO1(L)"            "EBr2r4"            "GNG"              
#>   [4] "vACA(R)"           "b'2(R)"            "NO2(R)"           
#>   [7] "LX(L)"             "bL(R)"             "PB(L4)"           
#>  [10] "FBl7"              "EBr6"              "SPS(L)"           
#>  [13] "PB(R1)"            "bL(L)"             "aL(R)"            
#>  [16] "PB(R7)"            "b'L(R)"            "PB"               
#>  [19] "SAD"               "NO1(R)"            "VES(R)"           
#>  [22] "FBl3"              "FBl4"              "WED(R)"           
#>  [25] "ATL(R)"            "RUB(L)"            "PB(L5)"           
#>  [28] "LX(R)"             "g1(R)"             "AOT(R)"           
#>  [31] "ATL(L)"            "mALT(R)"           "FBl5"             
#>  [34] "NO(L)"             "ROB(R)"            "PB(L6)"           
#>  [37] "FBl1"              "IPS(R)"            "SCL(L)"           
#>  [40] "PRW"               "EPA(R)"            "PB(L7)"           
#>  [43] "PB(R4)"            "CA(L)"             "PLP(R)"           
#>  [46] "ICL(R)"            "OL(R)"             "PENP"             
#>  [49] "INP"               "a'1(R)"            "GC"               
#>  [52] "PB(L1)"            "ICL(L)"            "LAL(-GA)(R)"      
#>  [55] "a'L(R)"            "g4(R)"             "NO"               
#>  [58] "NO(R)"             "PB(L9)"            "LH(R)"            
#>  [61] "CAN(R)"            "CRE(R)"            "IB"               
#>  [64] "a2(R)"             "SAD(-AMMC)"        "LAL(L)"           
#>  [67] "gL(R)"             "LO(R)"             "b1(R)"            
#>  [70] "a'L(L)"            "a1(R)"             "AOTU(R)"          
#>  [73] "a'2(R)"            "CRE(L)"            "LOP(R)"           
#>  [76] "AMMC"              "BU(L)"             "AB(R)"            
#>  [79] "b'L(L)"            "PB(R8)"            "lACA(R)"          
#>  [82] "SIP(R)"            "PVLP(R)"           "POC"              
#>  [85] "a'3(R)"            "SPS(R)"            "PB(R9)"           
#>  [88] "SNP(L)"            "FBl6"              "MB(R)"            
#>  [91] "GOR(L)"            "SCL(R)"            "EBr3pw"           
#>  [94] "MB(+ACA)(R)"       "AVLP(R)"           "RUB(R)"           
#>  [97] "PED(R)"            "EBr3d"             "FBl9"             
#> [100] "EPA(L)"            "GA(R)"             "gL(L)"            
#> [103] "g5(R)"             "AME(R)"            "b'1(R)"           
#> [106] "PB(L3)"            "GF(R)"             "ME(R)"            
#> [109] "EB"                "AB(L)"             "NO2(L)"           
#> [112] "PB(R6)"            "b2(R)"             "MB(L)"            
#> [115] "CA(R)"             "VES(L)"            "VLNP(R)"          
#> [118] "EBr3am"            "CRE(-ROB,-RUB)(R)" "NO3(R)"           
#> [121] "BU(R)"             "AL(L)"             "PB(R2)"           
#> [124] "VMNP"              "aL(L)"             "GOR(R)"           
#> [127] "PB(R3)"            "hemibrain"         "EBr1"             
#> [130] "PB(L8)"            "g3(R)"             "FLA(R)"           
#> [133] "SNP(R)"            "NO3(L)"            "FBl2"             
#> [136] "FBl8"              "LAL(R)"            "SLP(R)"           
#> [139] "FB"                "PB(L2)"            "mALT(L)"          
#> [142] "CX"                "SMP(R)"            "EBr5"             
#> [145] "dACA(R)"           "AL(R)"             "PB(R5)"           
#> [148] "g2(R)"             "SIP(L)"            "a3(R)"            
#> [151] "SMP(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-D(L)"           "LAL(R)"            "AL-DL2d(R)"       
#>   [4] "NO(L)"             "AL-VP5(R)"         "SMP(R)"           
#>   [7] "BU(R)"             "NO2(R)"            "RUB(L)"           
#>  [10] "a3(R)"             "LAL(L)"            "NO2(L)"           
#>  [13] "POC"               "SMP(L)"            "PRW"              
#>  [16] "AL-DM6(L)"         "AL-DM3(R)"         "CX"               
#>  [19] "NO(R)"             "GOR(R)"            "AL-VA1v(R)"       
#>  [22] "AOT(R)"            "gL(L)"             "aL(R)"            
#>  [25] "PB(R8)"            "PB"                "LX(R)"            
#>  [28] "AL-VC2(R)"         "AL-VP1m(R)"        "FLA(R)"           
#>  [31] "EB"                "ICL(R)"            "AL-DL1(R)"        
#>  [34] "AL-DM2(R)"         "AL-DP1m(R)"        "AL-DA3(L)"        
#>  [37] "AL-DP1m(L)"        "AB(R)"             "PB(L4)"           
#>  [40] "GF(R)"             "CRE(-RUB)(L)"      "PVLP(R)"          
#>  [43] "AL-D(R)"           "SIP(L)"            "FB"               
#>  [46] "a'2(R)"            "NO"                "FBl5"             
#>  [49] "AL-VA4(R)"         "CRE(L)"            "LH(R)"            
#>  [52] "EBr5"              "EBr3pw"            "FB-column3"       
#>  [55] "AL-VM1(R)"         "MB(R)"             "AL-VA7l(R)"       
#>  [58] "AL-VC1(R)"         "PB(R9)"            "FBl2"             
#>  [61] "GNG"               "g3(R)"             "ME(R)"            
#>  [64] "SNP(R)"            "g4(R)"             "a'3(R)"           
#>  [67] "AL-DL5(R)"         "AMMC"              "NO1(L)"           
#>  [70] "g2(R)"             "AL-DC4(L)"         "FBl8"             
#>  [73] "SPS(R)"            "AVLP(R)"           "AL-DA2(R)"        
#>  [76] "AL-VP2(R)"         "g5(R)"             "AL-DC3(R)"        
#>  [79] "RUB(R)"            "ATL(R)"            "LX(L)"            
#>  [82] "AL(R)"             "AL-DA2(L)"         "FBl3"             
#>  [85] "SLP(R)"            "AL-DL2v(R)"        "AL-V(R)"          
#>  [88] "a1(R)"             "SPS(L)"            "EPA(R)"           
#>  [91] "AL(L)"             "AL-VC5(R)"         "FBl1"             
#>  [94] "AL-DP1l(R)"        "PB(L7)"            "INP"              
#>  [97] "AL-DM2(L)"         "gL(R)"             "LAL(-GA)(R)"      
#> [100] "ROB(R)"            "AL-VP1l(R)"        "VMNP"             
#> [103] "AL-DC2(L)"         "AL-DM1(L)"         "MB(+ACA)(R)"      
#> [106] "PB(L5)"            "CRE(-ROB,-RUB)(R)" "AL-VA6(L)"        
#> [109] "FBl9"              "EPA(L)"            "AB(L)"            
#> [112] "GC"                "PLP(R)"            "FBl7"             
#> [115] "b1(R)"             "AL-DM4(R)"         "AL-VC3m(R)"       
#> [118] "CRE(R)"            "FBl4"              "AME(R)"           
#> [121] "AL-VA3(R)"         "AL-DM5(L)"         "AL-DA4m(L)"       
#> [124] "aL(L)"             "AL-DM5(R)"         "vACA(R)"          
#> [127] "PB(L3)"            "EBr1"              "AL-DL4(L)"        
#> [130] "AL-VM4(R)"         "AL-VM7v(L)"        "AL-VM7v(R)"       
#> [133] "EBr3am"            "AL-VL1(R)"         "AL-VP3(R)"        
#> [136] "SCL(L)"            "b'2(R)"            "AOTU(R)"          
#> [139] "NO1(R)"            "SIP(R)"            "AL-VA7m(R)"       
#> [142] "AL-VM5d(R)"        "PB(L9)"            "AL-VM2(R)"        
#> [145] "AL-VC3l(R)"        "AL-VL2p(R)"        "AL-VP1d(R)"       
#> [148] "EBr6"              "bL(L)"             "SAD(-AMMC)"       
#> [151] "bL(R)"             "a'1(R)"            "LO(R)"            
#> [154] "AL-VM5v(R)"        "PB(R6)"            "CA(R)"            
#> [157] "VES(L)"            "AL-DC4(R)"         "ATL(L)"           
#> [160] "GOR(L)"            "g1(R)"             "AL-DC1(L)"        
#> [163] "ICL(L)"            "mALT(R)"           "b'L(R)"           
#> [166] "AL-DM1(R)"         "AL-VL2a(R)"        "AL-VM3(R)"        
#> [169] "PENP"              "WED(R)"            "PB(L8)"           
#> [172] "AL-DA4l(R)"        "GA(R)"             "EBr3d"            
#> [175] "PB(R5)"            "b'1(R)"            "hemibrain"        
#> [178] "SNP(L)"            "VES(R)"            "PB(R7)"           
#> [181] "AL-DM3(L)"         "AL-DM4(L)"         "a'L(L)"           
#> [184] "AL-DA4m(R)"        "LOP(R)"            "BU(L)"            
#> [187] "AL-VM7d(L)"        "PB(R2)"            "IPS(R)"           
#> [190] "AL-DC1(R)"         "a2(R)"             "lACA(R)"          
#> [193] "PB(R1)"            "CAN(R)"            "AL-VA1d(R)"       
#> [196] "IB"                "AL-VP4(R)"         "CA(L)"            
#> [199] "OL(R)"             "AL-DL3(R)"         "AL-DL4(R)"        
#> [202] "AL-DA3(R)"         "AL-VA6(R)"         "dACA(R)"          
#> [205] "PB(L1)"            "AL-DL5(L)"         "NO3(L)"           
#> [208] "AL-VM7d(R)"        "SAD"               "MB(L)"            
#> [211] "SCL(R)"            "AL-VC4(R)"         "mALT(L)"          
#> [214] "PED(R)"            "PB(R3)"            "AL-DC2(R)"        
#> [217] "PB(L2)"            "AL-DA1(R)"         "EBr2r4"           
#> [220] "a'L(R)"            "FBl6"              "PB(R4)"           
#> [223] "b2(R)"             "b'L(L)"            "AL-VA5(R)"        
#> [226] "AL-DM6(R)"         "NO3(R)"            "AL-VA2(R)"        
#> [229] "PB(L6)"            "VLNP(R)"          
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-DA4m(L)"        "AL-DM2(L)"         "FBl2"             
#>   [4] "b1(R)"             "MB(L)"             "PLP(R)"           
#>   [7] "AL-DL2d(R)"        "AL-DM3(R)"         "b'L(L)"           
#>  [10] "AL-VC3m(R)"        "b'1(R)"            "a2(R)"            
#>  [13] "aL(R)"             "AL-DA3(L)"         "aL(L)"            
#>  [16] "AL-DL3(R)"         "PENP"              "lACA(R)"          
#>  [19] "AL-DM6(L)"         "AL-DL4(R)"         "a'3(R)"           
#>  [22] "AL-DC4(R)"         "IB"                "b2(R)"            
#>  [25] "SPS(L)"            "LX(R)"             "gL(L)"            
#>  [28] "GNG"               "PED(R)"            "AL-DA4l(R)"       
#>  [31] "AL-VA6(R)"         "VLNP(R)"           "CRE(R)"           
#>  [34] "AL-DC2(R)"         "NO"                "FLA(R)"           
#>  [37] "SNP(R)"            "AL-VM4(R)"         "SIP(L)"           
#>  [40] "EBr6"              "EBr2r4"            "g5(R)"            
#>  [43] "FBl1"              "NO(R)"             "SLP(R)"           
#>  [46] "GF(R)"             "b'L(R)"            "NO2(L)"           
#>  [49] "GOR(L)"            "WED(R)"            "BU(L)"            
#>  [52] "LH(R)"             "GOR(R)"            "AL-DC2(L)"        
#>  [55] "AL-VM2(R)"         "AB(L)"             "EBr3am"           
#>  [58] "VMNP"              "AL-VA6(L)"         "SIP(R)"           
#>  [61] "PB(L6)"            "RUB(R)"            "dACA(R)"          
#>  [64] "CAN(R)"            "LAL(R)"            "AL-DA3(R)"        
#>  [67] "AL-VA7m(R)"        "g1(R)"             "AL-VA4(R)"        
#>  [70] "ATL(L)"            "AL-DM4(L)"         "SNP(L)"           
#>  [73] "AL-DC1(R)"         "AL-VL2p(R)"        "AL-VP4(R)"        
#>  [76] "AL(L)"             "AL-DA2(L)"         "PB(L9)"           
#>  [79] "AL-D(R)"           "NO2(R)"            "AL-VA1d(R)"       
#>  [82] "AL-VM5d(R)"        "ROB(R)"            "EPA(L)"           
#>  [85] "AL-VC2(R)"         "AL-VM3(R)"         "CRE(-ROB,-RUB)(R)"
#>  [88] "AL-VP2(R)"         "mALT(R)"           "AL-D(L)"          
#>  [91] "FBl8"              "RUB(L)"            "AL-VC5(R)"        
#>  [94] "AL-DL5(L)"         "SAD(-AMMC)"        "PVLP(R)"          
#>  [97] "AL-VC4(R)"         "EB"                "AL-DC1(L)"        
#> [100] "CA(R)"             "AL-VL2a(R)"        "AL-VA1v(R)"       
#> [103] "AL-VM5v(R)"        "AL-VP1m(R)"        "SMP(L)"           
#> [106] "CRE(-RUB)(L)"      "VES(L)"            "PB(R5)"           
#> [109] "NO3(L)"            "AL-DA2(R)"         "AL-DP1m(R)"       
#> [112] "AL-DC3(R)"         "PB(L2)"            "hemibrain"        
#> [115] "PB(L7)"            "IPS(R)"            "AL-VA5(R)"        
#> [118] "AL-VM7d(R)"        "EPA(R)"            "AL-DC4(L)"        
#> [121] "FBl7"              "AL-DM2(R)"         "AL-DM4(R)"        
#> [124] "SPS(R)"            "MB(+ACA)(R)"       "FB-column3"       
#> [127] "OL(R)"             "AL-VA3(R)"         "EBr3d"            
#> [130] "EBr3pw"            "AME(R)"            "AMMC"             
#> [133] "FBl5"              "PB(R4)"            "bL(R)"            
#> [136] "AL(R)"             "g3(R)"             "FB"               
#> [139] "LAL(L)"            "SCL(L)"            "CA(L)"            
#> [142] "PB(L1)"            "PB(R8)"            "AL-DL4(L)"        
#> [145] "PB(R3)"            "CX"                "AL-VA7l(R)"       
#> [148] "PB(R6)"            "a3(R)"             "AL-DM1(R)"        
#> [151] "FBl3"              "FBl4"              "FBl9"             
#> [154] "AOT(R)"            "AL-VP3(R)"         "PB(L3)"           
#> [157] "LAL(-GA)(R)"       "AL-DP1m(L)"        "LX(L)"            
#> [160] "PB"                "AL-DL1(R)"         "AL-VA2(R)"        
#> [163] "AL-VP5(R)"         "mALT(L)"           "g4(R)"            
#> [166] "BU(R)"             "AL-DM5(L)"         "gL(R)"            
#> [169] "bL(L)"             "AL-V(R)"           "a1(R)"            
#> [172] "PRW"               "PB(R1)"            "AL-DA4m(R)"       
#> [175] "AL-DP1l(R)"        "POC"               "NO3(R)"           
#> [178] "NO1(L)"            "a'L(R)"            "AL-DM3(L)"        
#> [181] "AOTU(R)"           "FBl6"              "AVLP(R)"          
#> [184] "AL-DL2v(R)"        "AL-DM5(R)"         "a'1(R)"           
#> [187] "ATL(R)"            "b'2(R)"            "g2(R)"            
#> [190] "SCL(R)"            "PB(L8)"            "a'2(R)"           
#> [193] "AL-VM7v(R)"        "AL-VP1l(R)"        "PB(R9)"           
#> [196] "ICL(R)"            "INP"               "PB(L4)"           
#> [199] "a'L(L)"            "GA(R)"             "GC"               
#> [202] "LO(R)"             "VES(R)"            "PB(R2)"           
#> [205] "AB(R)"             "AL-VM7d(L)"        "CRE(L)"           
#> [208] "NO(L)"             "AL-VC1(R)"         "EBr5"             
#> [211] "ICL(L)"            "AL-DA1(R)"         "PB(R7)"           
#> [214] "AL-VL1(R)"         "LOP(R)"            "SAD"              
#> [217] "NO1(R)"            "AL-DM1(L)"         "MB(R)"            
#> [220] "AL-VC3l(R)"        "AL-VP1d(R)"        "vACA(R)"          
#> [223] "PB(L5)"            "EBr1"              "AL-DL5(R)"        
#> [226] "AL-DM6(R)"         "AL-VM1(R)"         "SMP(R)"           
#> [229] "ME(R)"             "AL-VM7v(L)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method   path                                     
#> [1,] "GET"    "/api/custom/custom"                     
#> [2,] "GET"    "/api/npexplorer/roisinneuron"           
#> [3,] "POST"   "/api/npexplorer/autapses"               
#> [4,] "DELETE" "/api/*"                                 
#> [5,] "POST"   "/api/v:ver/npexplorer/simpleconnections"
#> [6,] "GET"    "/api/v:ver/npexplorer/autapses"         
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"             
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"    
#> [4,] "github.com/labstack/echo/v4.glob..func1"                                                   
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }