Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "PRW"               "AL-VA3(R)"         "EBr6"             
#>   [4] "AL-DM2(L)"         "PB(R1)"            "VES(R)"           
#>   [7] "FBl9"              "IPS(R)"            "AL-D(R)"          
#>  [10] "a'L(R)"            "NO(L)"             "aL(R)"            
#>  [13] "BU(R)"             "AL-D(L)"           "gL(R)"            
#>  [16] "PB(L4)"            "AL-DC2(R)"         "LAL(L)"           
#>  [19] "SCL(R)"            "AL-DA4l(R)"        "RUB(R)"           
#>  [22] "SAD"               "PB(R2)"            "b'L(R)"           
#>  [25] "g5(R)"             "AL-DL1(R)"         "AL-VA6(R)"        
#>  [28] "AL-VM7d(L)"        "LX(L)"             "AL-DL5(R)"        
#>  [31] "SMP(R)"            "AMMC"              "FLA(R)"           
#>  [34] "PB(L9)"            "AL-DC3(R)"         "AL-VP1m(R)"       
#>  [37] "AL-VP4(R)"         "EBr3am"            "PB(L5)"           
#>  [40] "PB(R7)"            "a'1(R)"            "aL(L)"            
#>  [43] "AL-VM2(R)"         "AL-VP1l(R)"        "EBr2r4"           
#>  [46] "OL(R)"             "AL-DA3(R)"         "AL-DM4(R)"        
#>  [49] "AL-VA1v(R)"        "AL-VC3m(R)"        "VLNP(R)"          
#>  [52] "LO(R)"             "NO2(R)"            "GNG"              
#>  [55] "AL-DP1m(R)"        "GOR(L)"            "PB(L1)"           
#>  [58] "LX(R)"             "FBl2"              "AL-VM3(R)"        
#>  [61] "AL-DL2d(R)"        "FBl4"              "AL-DM3(R)"        
#>  [64] "AL-VL2p(R)"        "SPS(R)"            "FBl5"             
#>  [67] "b1(R)"             "PB(R4)"            "AL-DM1(R)"        
#>  [70] "AL-VL1(R)"         "CRE(R)"            "AL-DA3(L)"        
#>  [73] "AL(R)"             "AL-DM2(R)"         "g2(R)"            
#>  [76] "a'2(R)"            "AL-VA1d(R)"        "IB"               
#>  [79] "NO3(L)"            "ATL(R)"            "AL-DM5(R)"        
#>  [82] "CA(L)"             "FB-column3"        "LAL(-GA)(R)"      
#>  [85] "mALT(R)"           "GC"                "WED(R)"           
#>  [88] "AL-DA2(L)"         "AL-DA1(R)"         "mALT(L)"          
#>  [91] "b'2(R)"            "EBr5"              "NO"               
#>  [94] "AVLP(R)"           "ICL(R)"            "AL-VP1d(R)"       
#>  [97] "AB(L)"             "AL-DC1(L)"         "FBl3"             
#> [100] "PB(R3)"            "SMP(L)"            "PB(L3)"           
#> [103] "PED(R)"            "CA(R)"             "AL-DA4m(R)"       
#> [106] "AL-VC1(R)"         "AL-V(R)"           "b'1(R)"           
#> [109] "PB"                "AL-DL5(L)"         "FBl8"             
#> [112] "AME(R)"            "AL-VC3l(R)"        "AL-VL2a(R)"       
#> [115] "AL-VM7d(R)"        "PB(R5)"            "BU(L)"            
#> [118] "g4(R)"             "EPA(R)"            "AL-VM5d(R)"       
#> [121] "CX"                "LAL(R)"            "hemibrain"        
#> [124] "dACA(R)"           "AL(L)"             "ROB(R)"           
#> [127] "AL-DC1(R)"         "POC"               "FB"               
#> [130] "gL(L)"             "AL-DP1m(L)"        "AL-DM6(R)"        
#> [133] "AL-VA2(R)"         "AL-DM4(L)"         "AOTU(R)"          
#> [136] "SIP(L)"            "VES(L)"            "AL-VM1(R)"        
#> [139] "AL-DL4(L)"         "AL-VA5(R)"         "MB(L)"            
#> [142] "FBl6"              "SLP(R)"            "AL-VM5v(R)"       
#> [145] "EB"                "SIP(R)"            "PB(L2)"           
#> [148] "a'3(R)"            "AL-VA6(L)"         "PB(R6)"           
#> [151] "AL-DA2(R)"         "AL-VC2(R)"         "AOT(R)"           
#> [154] "vACA(R)"           "AL-DM1(L)"         "AL-VA4(R)"        
#> [157] "AL-VA7m(R)"        "SCL(L)"            "lACA(R)"          
#> [160] "AL-DM3(L)"         "FBl1"              "AL-DL3(R)"        
#> [163] "CRE(-ROB,-RUB)(R)" "AL-DC4(L)"         "AL-VM7v(L)"       
#> [166] "CRE(-RUB)(L)"      "bL(L)"             "ICL(L)"           
#> [169] "LOP(R)"            "a1(R)"             "b2(R)"            
#> [172] "PB(L6)"            "NO1(L)"            "PLP(R)"           
#> [175] "AL-DL2v(R)"        "NO2(L)"            "AL-VA7l(R)"       
#> [178] "SAD(-AMMC)"        "LH(R)"             "EBr1"             
#> [181] "IB112"             "g3(R)"             "AL-DC2(L)"        
#> [184] "NO(R)"             "FBl7"              "AL-VM7v(R)"       
#> [187] "AL-DM5(L)"         "AB(R)"             "CAN(R)"           
#> [190] "INP"               "bL(R)"             "GA(R)"            
#> [193] "a3(R)"             "b'L(L)"            "AL-VC4(R)"        
#> [196] "g1(R)"             "MB(+ACA)(R)"       "AL-DL4(R)"        
#> [199] "PB(R9)"            "EPA(L)"            "a2(R)"            
#> [202] "PB(L8)"            "SNP(L)"            "PVLP(R)"          
#> [205] "PB(R8)"            "EBr3d"             "SNP(R)"           
#> [208] "EBr3pw"            "a'L(L)"            "GF(R)"            
#> [211] "CRE(L)"            "PB(L7)"            "AL-VM4(R)"        
#> [214] "AL-VP3(R)"         "NO1(R)"            "SPS(L)"           
#> [217] "AL-DA4m(L)"        "AL-DC4(R)"         "AL-VC5(R)"        
#> [220] "ATL(L)"            "ME(R)"             "AL-DM6(L)"        
#> [223] "RUB(L)"            "MB(R)"             "AL-VP5(R)"        
#> [226] "PENP"              "NO3(R)"            "VMNP"             
#> [229] "GOR(R)"            "AL-DP1l(R)"        "AL-VP2(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "CA(R)"             "IB"                "SCL(L)"           
#>   [4] "AMMC"              "SNP(R)"            "gL(L)"            
#>   [7] "SNP(L)"            "vACA(R)"           "b'2(R)"           
#>  [10] "VLNP(R)"           "NO3(R)"            "PB(R2)"           
#>  [13] "CAN(R)"            "PB(R6)"            "g5(R)"            
#>  [16] "OL(R)"             "a'3(R)"            "PB(R7)"           
#>  [19] "SIP(L)"            "PED(R)"            "PB(L7)"           
#>  [22] "AB(L)"             "CA(L)"             "a2(R)"            
#>  [25] "PRW"               "VMNP"              "g4(R)"            
#>  [28] "FBl1"              "AB(R)"             "SCL(R)"           
#>  [31] "PB(R1)"            "LAL(R)"            "MB(R)"            
#>  [34] "PB(R8)"            "NO3(L)"            "ICL(R)"           
#>  [37] "POC"               "PB(R5)"            "b'1(R)"           
#>  [40] "GF(R)"             "CRE(R)"            "FBl9"             
#>  [43] "aL(L)"             "RUB(R)"            "ATL(L)"           
#>  [46] "SAD"               "dACA(R)"           "EPA(R)"           
#>  [49] "CRE(L)"            "FBl2"              "LX(L)"            
#>  [52] "g1(R)"             "FLA(R)"            "VES(R)"           
#>  [55] "EBr3am"            "PB(L5)"            "GOR(L)"           
#>  [58] "ICL(L)"            "FBl3"              "FBl5"             
#>  [61] "b'L(L)"            "CX"                "PENP"             
#>  [64] "PB(L3)"            "EBr6"              "GOR(R)"           
#>  [67] "SMP(R)"            "mALT(L)"           "GA(R)"            
#>  [70] "lACA(R)"           "gL(R)"             "PB(L4)"           
#>  [73] "LOP(R)"            "CRE(-ROB,-RUB)(R)" "mALT(R)"          
#>  [76] "EBr2r4"            "LAL(L)"            "GC"               
#>  [79] "PB(L1)"            "PLP(R)"            "EB"               
#>  [82] "ROB(R)"            "EBr3pw"            "FBl6"             
#>  [85] "b1(R)"             "a1(R)"             "SPS(R)"           
#>  [88] "SIP(R)"            "NO"                "a'1(R)"           
#>  [91] "IPS(R)"            "PB(L2)"            "WED(R)"           
#>  [94] "g2(R)"             "SMP(L)"            "g3(R)"            
#>  [97] "NO1(L)"            "BU(L)"             "NO1(R)"           
#> [100] "PB"                "ME(R)"             "EBr3d"            
#> [103] "a'L(R)"            "PVLP(R)"           "ATL(R)"           
#> [106] "AL(R)"             "aL(R)"             "b2(R)"            
#> [109] "AME(R)"            "NO2(R)"            "hemibrain"        
#> [112] "PB(L6)"            "MB(+ACA)(R)"       "EPA(L)"           
#> [115] "VES(L)"            "bL(L)"             "RUB(L)"           
#> [118] "PB(R4)"            "FB"                "PB(L8)"           
#> [121] "PB(L9)"            "INP"               "a'L(L)"           
#> [124] "SLP(R)"            "LAL(-GA)(R)"       "b'L(R)"           
#> [127] "a'2(R)"            "FBl4"              "FBl8"             
#> [130] "AVLP(R)"           "NO2(L)"            "SPS(L)"           
#> [133] "LX(R)"             "PB(R9)"            "NO(R)"            
#> [136] "LH(R)"             "GNG"               "a3(R)"            
#> [139] "AOT(R)"            "MB(L)"             "BU(R)"            
#> [142] "AL(L)"             "bL(R)"             "EBr1"             
#> [145] "EBr5"              "PB(R3)"            "AOTU(R)"          
#> [148] "SAD(-AMMC)"        "LO(R)"             "FBl7"             
#> [151] "NO(L)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "SAD"               "dACA(R)"           "AL-DC2(L)"        
#>   [4] "AL-DL5(L)"         "PB(L7)"            "ICL(R)"           
#>   [7] "bL(R)"             "AL-VM4(R)"         "AOTU(R)"          
#>  [10] "AL-VL2p(R)"        "RUB(R)"            "a2(R)"            
#>  [13] "NO(R)"             "NO3(L)"            "AL-VC5(R)"        
#>  [16] "AOT(R)"            "MB(L)"             "AL-DA4m(L)"       
#>  [19] "AL-VA7m(R)"        "PB(L2)"            "PB(L6)"           
#>  [22] "AL-DA4l(R)"        "AL-DM1(R)"         "AL-DM6(R)"        
#>  [25] "AL-VA6(R)"         "SPS(L)"            "IB"               
#>  [28] "PB(L5)"            "SIP(L)"            "PB(L9)"           
#>  [31] "b'L(L)"            "SMP(L)"            "AL-DM5(R)"        
#>  [34] "AL-VL1(R)"         "NO3(R)"            "ME(R)"            
#>  [37] "SNP(R)"            "PB(R4)"            "AL-DC2(R)"        
#>  [40] "AL-DC1(L)"         "AMMC"              "NO1(L)"           
#>  [43] "AL-DC1(R)"         "AL-DC3(R)"         "CA(R)"            
#>  [46] "MB(R)"             "VLNP(R)"           "LOP(R)"           
#>  [49] "CX"                "AL-DL1(R)"         "g3(R)"            
#>  [52] "GOR(L)"            "LX(R)"             "SAD(-AMMC)"       
#>  [55] "PB(R2)"            "PB"                "INP"              
#>  [58] "AL-V(R)"           "AB(L)"             "NO"               
#>  [61] "PB(R3)"            "NO2(R)"            "PB(R8)"           
#>  [64] "AL-DL4(L)"         "LX(L)"             "AL-VM5v(R)"       
#>  [67] "CRE(-ROB,-RUB)(R)" "AL-DL2d(R)"        "POC"              
#>  [70] "GA(R)"             "NO(L)"             "AL-VC3m(R)"       
#>  [73] "EBr2r4"            "AL-DA4m(R)"        "SCL(R)"           
#>  [76] "AL-DM4(L)"         "ATL(R)"            "AL-VA1v(R)"       
#>  [79] "AL-VM7v(R)"        "EBr3pw"            "SNP(L)"           
#>  [82] "RUB(L)"            "GNG"               "AL-VA4(R)"        
#>  [85] "PB(R7)"            "EB"                "b2(R)"            
#>  [88] "AL-D(L)"           "AL-DA1(R)"         "b'2(R)"           
#>  [91] "PB(L3)"            "ICL(L)"            "OL(R)"            
#>  [94] "LAL(L)"            "LAL(-GA)(R)"       "BU(L)"            
#>  [97] "PB(L8)"            "a'2(R)"            "FB"               
#> [100] "gL(L)"             "NO1(R)"            "a'1(R)"           
#> [103] "IPS(R)"            "PENP"              "GC"               
#> [106] "AL-DL3(R)"         "AL-VM5d(R)"        "WED(R)"           
#> [109] "AL-DL4(R)"         "EBr3d"             "FBl2"             
#> [112] "AL-DC4(L)"         "AL-VM7d(L)"        "AL-VA3(R)"        
#> [115] "GF(R)"             "PED(R)"            "AB(R)"            
#> [118] "CA(L)"             "g2(R)"             "SIP(R)"           
#> [121] "MB(+ACA)(R)"       "FBl8"              "GOR(R)"           
#> [124] "hemibrain"         "AL-VA6(L)"         "AL(R)"            
#> [127] "AL-DM4(R)"         "FBl1"              "LAL(R)"           
#> [130] "AL-DL2v(R)"        "CRE(R)"            "EBr5"             
#> [133] "PLP(R)"            "AL-DM1(L)"         "AL-VM7v(L)"       
#> [136] "EBr3am"            "aL(R)"             "PB(R9)"           
#> [139] "FBl9"              "AL-VC2(R)"         "SMP(R)"           
#> [142] "AL-DA2(R)"         "g5(R)"             "FB-column3"       
#> [145] "AL-DM2(R)"         "gL(R)"             "bL(L)"            
#> [148] "AL-VA1d(R)"        "AL-VP5(R)"         "AL-VP3(R)"        
#> [151] "EBr6"              "VMNP"              "BU(R)"            
#> [154] "FBl3"              "FBl4"              "AL-VP4(R)"        
#> [157] "AL-VP1m(R)"        "a1(R)"             "PRW"              
#> [160] "AVLP(R)"           "AL-DC4(R)"         "AL-VC3l(R)"       
#> [163] "AL-VL2a(R)"        "PB(R6)"            "SLP(R)"           
#> [166] "AL-DL5(R)"         "a'3(R)"            "PB(R1)"           
#> [169] "aL(L)"             "AL-DP1l(R)"        "AL-VM1(R)"        
#> [172] "PB(R5)"            "EPA(R)"            "b'L(R)"           
#> [175] "EBr1"              "ROB(R)"            "PVLP(R)"          
#> [178] "g1(R)"             "AL-DP1m(R)"        "AL-VM2(R)"        
#> [181] "AL-DM5(L)"         "FBl6"              "AL-DM3(R)"        
#> [184] "AL-DM2(L)"         "FBl7"              "AL-VA5(R)"        
#> [187] "CAN(R)"            "AL-VC1(R)"         "AL-VM3(R)"        
#> [190] "lACA(R)"           "AL(L)"             "AL-DP1m(L)"       
#> [193] "vACA(R)"           "AL-DA3(L)"         "LH(R)"            
#> [196] "a'L(L)"            "AL-VP2(R)"         "PB(L1)"           
#> [199] "CRE(-RUB)(L)"      "VES(R)"            "VES(L)"           
#> [202] "AL-DM6(L)"         "FBl5"              "AL-D(R)"          
#> [205] "AL-VM7d(R)"        "mALT(L)"           "a'L(R)"           
#> [208] "EPA(L)"            "a3(R)"             "AL-VA7l(R)"       
#> [211] "AL-VP1d(R)"        "SPS(R)"            "g4(R)"            
#> [214] "CRE(L)"            "b1(R)"             "SCL(L)"           
#> [217] "FLA(R)"            "AME(R)"            "AL-DA3(R)"        
#> [220] "AL-VC4(R)"         "LO(R)"             "AL-VA2(R)"        
#> [223] "NO2(L)"            "AL-DA2(L)"         "mALT(R)"          
#> [226] "PB(L4)"            "b'1(R)"            "AL-DM3(L)"        
#> [229] "AL-VP1l(R)"        "ATL(L)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "EBr1"              "SLP(R)"            "AL-VA3(R)"        
#>   [4] "EBr2r4"            "AL-DA3(L)"         "AL-DM4(L)"        
#>   [7] "GOR(L)"            "g1(R)"             "EBr6"             
#>  [10] "a'3(R)"            "CAN(R)"            "MB(R)"            
#>  [13] "AL-DM1(R)"         "AL-VA7m(R)"        "AL-VP5(R)"        
#>  [16] "AL-VP1l(R)"        "NO1(R)"            "PB(L7)"           
#>  [19] "AL-DM5(R)"         "a3(R)"             "AL-VL2p(R)"       
#>  [22] "vACA(R)"           "SIP(L)"            "AL-VA6(L)"        
#>  [25] "NO"                "b'L(R)"            "GNG"              
#>  [28] "AL-VA2(R)"         "AL-DL5(L)"         "NO(R)"            
#>  [31] "AL-VA1v(R)"        "PB(R9)"            "aL(L)"            
#>  [34] "AL-DL5(R)"         "VMNP"              "PB(R2)"           
#>  [37] "AL-VA4(R)"         "AL-VP3(R)"         "PB(L5)"           
#>  [40] "EPA(R)"            "AL-VA5(R)"         "AL-VC3l(R)"       
#>  [43] "AL-DM1(L)"         "AL-DM2(R)"         "CX"               
#>  [46] "VLNP(R)"           "ICL(L)"            "NO(L)"            
#>  [49] "AL-DC4(R)"         "AL(R)"             "AL-VM7d(R)"       
#>  [52] "EBr5"              "a'L(R)"            "FBl6"             
#>  [55] "EBr3pw"            "AL-DL2d(R)"        "AL-VM1(R)"        
#>  [58] "FB-column3"        "PB(R5)"            "CA(L)"            
#>  [61] "AB(L)"             "b'2(R)"            "GA(R)"            
#>  [64] "PB(L1)"            "SPS(R)"            "SNP(L)"           
#>  [67] "AL-DP1m(L)"        "PB(L6)"            "FBl7"             
#>  [70] "AL-VM5d(R)"        "EBr3am"            "PB(R7)"           
#>  [73] "a'2(R)"            "AL-DM5(L)"         "ME(R)"            
#>  [76] "AL-DC1(L)"         "FBl3"              "PB(L4)"           
#>  [79] "AL-DM4(R)"         "lACA(R)"           "LAL(-GA)(R)"      
#>  [82] "mALT(R)"           "AL-VA1d(R)"        "AL-VL1(R)"        
#>  [85] "SAD"               "FBl9"              "a1(R)"            
#>  [88] "g4(R)"             "MB(+ACA)(R)"       "MB(L)"            
#>  [91] "PB(L8)"            "PB(R8)"            "AL-VM4(R)"        
#>  [94] "WED(R)"            "AL-VM7d(L)"        "FBl1"             
#>  [97] "AL-VP1m(R)"        "CRE(R)"            "EB"               
#> [100] "AL-DP1m(R)"        "a2(R)"             "BU(R)"            
#> [103] "AL-VM7v(R)"        "SMP(R)"            "mALT(L)"          
#> [106] "GF(R)"             "SNP(R)"            "b'L(L)"           
#> [109] "AL-VC4(R)"         "SIP(R)"            "b1(R)"            
#> [112] "AL-DA4l(R)"        "NO3(L)"            "INP"              
#> [115] "SCL(L)"            "AL-DA4m(L)"        "NO2(R)"           
#> [118] "AL-DC3(R)"         "PB(R6)"            "SPS(L)"           
#> [121] "AL-VC1(R)"         "NO1(L)"            "AL-DC2(R)"        
#> [124] "AL-VC5(R)"         "LH(R)"             "AL-VP1d(R)"       
#> [127] "LAL(R)"            "AL-DM6(R)"         "BU(L)"            
#> [130] "FLA(R)"            "a'1(R)"            "PVLP(R)"          
#> [133] "AL-VM3(R)"         "AL-VM5v(R)"        "PB(L3)"           
#> [136] "AL-VL2a(R)"        "PB(R4)"            "AL-DA2(R)"        
#> [139] "g2(R)"             "AL-DL4(L)"         "ICL(R)"           
#> [142] "AL-DC4(L)"         "AOTU(R)"           "AL-DA2(L)"        
#> [145] "FBl2"              "g3(R)"             "AL-DM2(L)"        
#> [148] "VES(R)"            "AL-VP4(R)"         "PB"               
#> [151] "g5(R)"             "LO(R)"             "AVLP(R)"          
#> [154] "AL-DP1l(R)"        "AL-VC2(R)"         "RUB(L)"           
#> [157] "AL-DL3(R)"         "AL-VC3m(R)"        "FB"               
#> [160] "ATL(R)"            "IPS(R)"            "LX(R)"            
#> [163] "AL-DC2(L)"         "CA(R)"             "AL-DM3(L)"        
#> [166] "gL(R)"             "AL-DL4(R)"         "AL-DC1(R)"        
#> [169] "AL-VA6(R)"         "AL-DM3(R)"         "aL(R)"            
#> [172] "PB(R1)"            "AB(R)"             "bL(L)"            
#> [175] "AL-V(R)"           "IB"                "dACA(R)"          
#> [178] "CRE(L)"            "AL-VM2(R)"         "AL-DL1(R)"        
#> [181] "LAL(L)"            "b2(R)"             "AMMC"             
#> [184] "PENP"              "AL-DA1(R)"         "b'1(R)"           
#> [187] "AL(L)"             "CRE(-RUB)(L)"      "bL(R)"            
#> [190] "FBl8"              "OL(R)"             "RUB(R)"           
#> [193] "PB(L2)"            "SMP(L)"            "PRW"              
#> [196] "FBl4"              "AL-VA7l(R)"        "FBl5"             
#> [199] "CRE(-ROB,-RUB)(R)" "NO2(L)"            "hemibrain"        
#> [202] "AL-DM6(L)"         "ROB(R)"            "AL-DL2v(R)"       
#> [205] "AL-VP2(R)"         "gL(L)"             "EPA(L)"           
#> [208] "VES(L)"            "ATL(L)"            "PED(R)"           
#> [211] "a'L(L)"            "POC"               "SAD(-AMMC)"       
#> [214] "LX(L)"             "GOR(R)"            "AL-DA3(R)"        
#> [217] "AL-D(R)"           "NO3(R)"            "SCL(R)"           
#> [220] "EBr3d"             "PB(L9)"            "LOP(R)"           
#> [223] "GC"                "PLP(R)"            "AL-VM7v(L)"       
#> [226] "AOT(R)"            "AL-DA4m(R)"        "AL-D(L)"          
#> [229] "PB(R3)"            "AME(R)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method    path                                     
#> [1,] "POST"    "/api/npexplorer/celltype/:dataset/:type"
#> [2,] "GET"     "/api/available"                         
#> [3,] "CONNECT" "/api/*"                                 
#> [4,] "GET"     "/api/*"                                 
#> [5,] "GET"     "/api/npexplorer/findneurons"            
#> [6,] "GET"     "/api/dbmeta/version"                    
#>      name                                                                                       
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCellType-fm"   
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.SetupRoutes.func1"                        
#> [3,] "github.com/labstack/echo.glob..func1"                                                     
#> [4,] "github.com/labstack/echo.glob..func1"                                                     
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getFindNeurons-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getVersion-fm"            
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }