Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "g2(R)"             "AL-DL2v(R)"        "a1(R)"            
#>   [4] "ME(R)"             "CA(R)"             "AL-DA4m(R)"       
#>   [7] "AL-DC2(R)"         "PRW"               "EBr3pw"           
#>  [10] "FB-column3"        "AL-DL1(R)"         "AL-VL2p(R)"       
#>  [13] "EPA(L)"            "AL-VP1m(R)"        "MB(L)"            
#>  [16] "AL-DA4m(L)"        "AL-DM3(L)"         "NO3(L)"           
#>  [19] "PB(L7)"            "aL(L)"             "b'1(R)"           
#>  [22] "AL-VA1v(R)"        "IB"                "SCL(L)"           
#>  [25] "EBr2r4"            "CRE(L)"            "FBl2"             
#>  [28] "AL-DL3(R)"         "AL-VA4(R)"         "AL-VL1(R)"        
#>  [31] "PB(L8)"            "NO2(R)"            "b'L(L)"           
#>  [34] "SMP(L)"            "GA(R)"             "AL(L)"            
#>  [37] "AL-DM5(R)"         "GF(R)"             "g3(R)"            
#>  [40] "SIP(R)"            "VLNP(R)"           "AB(L)"            
#>  [43] "VMNP"              "AL(R)"             "AOT(R)"           
#>  [46] "AMMC"              "LAL(L)"            "hemibrain"        
#>  [49] "IB112"             "AL-DC4(R)"         "AL-DL2d(R)"       
#>  [52] "AL-VC3l(R)"        "CRE(-RUB)(L)"      "AL-VA6(R)"        
#>  [55] "EBr3am"            "LOP(R)"            "CRE(R)"           
#>  [58] "EB"                "NO(R)"             "ATL(R)"           
#>  [61] "PB(L5)"            "PB(L6)"            "AB(R)"            
#>  [64] "LX(L)"             "AL-VM7d(R)"        "ATL(L)"           
#>  [67] "PB(L4)"            "SCL(R)"            "AME(R)"           
#>  [70] "AL-D(R)"           "AL-VC3m(R)"        "AL-VP1l(R)"       
#>  [73] "PB(R5)"            "AL-DP1l(R)"        "BU(L)"            
#>  [76] "SNP(R)"            "AL-DP1m(L)"        "AL-DC1(R)"        
#>  [79] "AL-VA7m(R)"        "AL-VM3(R)"         "AL-DC4(L)"        
#>  [82] "AL-DM2(R)"         "AL-VM5v(R)"        "AL-V(R)"          
#>  [85] "SIP(L)"            "SAD(-AMMC)"        "ICL(R)"           
#>  [88] "EBr6"              "INP"               "AL-VC1(R)"        
#>  [91] "AL-VM5d(R)"        "GC"                "EBr3d"            
#>  [94] "AL-VM7d(L)"        "ICL(L)"            "AL-VA7l(R)"       
#>  [97] "AL-VL2a(R)"        "AL-DM1(L)"         "PB(R1)"           
#> [100] "GOR(R)"            "AL-VC4(R)"         "AL-VC2(R)"        
#> [103] "PVLP(R)"           "AL-DA4l(R)"        "AL-DL4(R)"        
#> [106] "b'L(R)"            "NO1(R)"            "gL(R)"            
#> [109] "LO(R)"             "MB(+ACA)(R)"       "GOR(L)"           
#> [112] "AL-DA3(L)"         "AL-DM5(L)"         "AVLP(R)"          
#> [115] "AL-VP5(R)"         "SMP(R)"            "AL-DM4(L)"        
#> [118] "AL-DM6(L)"         "LH(R)"             "AL-VM7v(R)"       
#> [121] "PENP"              "mALT(R)"           "a'L(R)"           
#> [124] "FBl4"              "g4(R)"             "AL-DM6(R)"        
#> [127] "b2(R)"             "AL-DM2(L)"         "g1(R)"            
#> [130] "PB(L9)"            "AL-DL5(R)"         "SAD"              
#> [133] "AL-D(L)"           "VES(L)"            "POC"              
#> [136] "NO2(L)"            "PB(R7)"            "VES(R)"           
#> [139] "FBl1"              "AL-VP3(R)"         "AL-VP1d(R)"       
#> [142] "CRE(-ROB,-RUB)(R)" "NO3(R)"            "AL-DC1(L)"        
#> [145] "AL-DC2(L)"         "SNP(L)"            "PB(R2)"           
#> [148] "AL-VM2(R)"         "lACA(R)"           "SPS(L)"           
#> [151] "MB(R)"             "CA(L)"             "PB(L3)"           
#> [154] "LAL(-GA)(R)"       "FBl8"              "AL-DA1(R)"        
#> [157] "AL-VA3(R)"         "AL-DA2(L)"         "a'1(R)"           
#> [160] "bL(L)"             "PB(R4)"            "bL(R)"            
#> [163] "NO"                "IPS(R)"            "a3(R)"            
#> [166] "AL-DM4(R)"         "AL-VM4(R)"         "PB"               
#> [169] "a'L(L)"            "AL-DA2(R)"         "AL-VP2(R)"        
#> [172] "a2(R)"             "dACA(R)"           "SLP(R)"           
#> [175] "vACA(R)"           "FB"                "EBr5"             
#> [178] "AL-VM7v(L)"        "FBl7"              "FBl9"             
#> [181] "CAN(R)"            "AL-DP1m(R)"        "AL-VA2(R)"        
#> [184] "RUB(R)"            "AOTU(R)"           "EBr1"             
#> [187] "b1(R)"             "AL-DM1(R)"         "CX"               
#> [190] "BU(R)"             "AL-DL4(L)"         "g5(R)"            
#> [193] "AL-DM3(R)"         "b'2(R)"            "OL(R)"            
#> [196] "PB(L2)"            "PED(R)"            "AL-VA6(L)"        
#> [199] "a'2(R)"            "EPA(R)"            "LX(R)"            
#> [202] "PB(R9)"            "FBl5"              "AL-DC3(R)"        
#> [205] "AL-VC5(R)"         "mALT(L)"           "SPS(R)"           
#> [208] "ROB(R)"            "LAL(R)"            "AL-VA1d(R)"       
#> [211] "PLP(R)"            "AL-DA3(R)"         "AL-VA5(R)"        
#> [214] "PB(R8)"            "FBl6"              "AL-VM1(R)"        
#> [217] "PB(R6)"            "WED(R)"            "RUB(L)"           
#> [220] "NO(L)"             "AL-VP4(R)"         "PB(L1)"           
#> [223] "AL-DL5(L)"         "GNG"               "a'3(R)"           
#> [226] "aL(R)"             "NO1(L)"            "FLA(R)"           
#> [229] "gL(L)"             "FBl3"              "PB(R3)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "SMP(L)"            "CRE(R)"            "FBl2"             
#>   [4] "FBl1"              "b'L(L)"            "a'2(R)"           
#>   [7] "gL(L)"             "PB(R3)"            "INP"              
#>  [10] "hemibrain"         "a'L(R)"            "PB(L6)"           
#>  [13] "bL(R)"             "GOR(R)"            "LAL(R)"           
#>  [16] "a'3(R)"            "a2(R)"             "SIP(L)"           
#>  [19] "PLP(R)"            "EB"                "FBl7"             
#>  [22] "AB(R)"             "AL(R)"             "AOT(R)"           
#>  [25] "EBr5"              "BU(L)"             "SNP(R)"           
#>  [28] "FBl4"              "IPS(R)"            "SLP(R)"           
#>  [31] "MB(R)"             "vACA(R)"           "aL(R)"            
#>  [34] "RUB(L)"            "GNG"               "mALT(L)"          
#>  [37] "BU(R)"             "PB(R4)"            "a1(R)"            
#>  [40] "ME(R)"             "bL(L)"             "PB(L7)"           
#>  [43] "b1(R)"             "EPA(L)"            "POC"              
#>  [46] "b'1(R)"            "PED(R)"            "PB(R8)"           
#>  [49] "CAN(R)"            "SCL(L)"            "PRW"              
#>  [52] "WED(R)"            "g1(R)"             "SCL(R)"           
#>  [55] "NO3(L)"            "g5(R)"             "CRE(-ROB,-RUB)(R)"
#>  [58] "PB(L3)"            "a'1(R)"            "a'L(L)"           
#>  [61] "EBr3am"            "IB"                "MB(L)"            
#>  [64] "NO1(R)"            "LX(R)"             "LO(R)"            
#>  [67] "MB(+ACA)(R)"       "LH(R)"             "AME(R)"           
#>  [70] "AMMC"              "PB(L9)"            "LOP(R)"           
#>  [73] "VLNP(R)"           "FBl6"              "PB(L4)"           
#>  [76] "SAD"               "NO1(L)"            "VES(R)"           
#>  [79] "PB(R5)"            "LAL(L)"            "AOTU(R)"          
#>  [82] "SPS(L)"            "g4(R)"             "OL(R)"            
#>  [85] "ATL(R)"            "PENP"              "GF(R)"            
#>  [88] "g3(R)"             "GA(R)"             "LAL(-GA)(R)"      
#>  [91] "EBr1"              "PB(L8)"            "ICL(R)"           
#>  [94] "LX(L)"             "FB"                "EBr2r4"           
#>  [97] "NO(R)"             "SAD(-AMMC)"        "PB(L2)"           
#> [100] "ATL(L)"            "SNP(L)"            "FBl3"             
#> [103] "NO2(R)"            "CX"                "EBr3pw"           
#> [106] "b'2(R)"            "PB(R7)"            "mALT(R)"          
#> [109] "GC"                "AL(L)"             "PB(R1)"           
#> [112] "gL(R)"             "CA(R)"             "PB(R2)"           
#> [115] "FBl9"              "AVLP(R)"           "CA(L)"            
#> [118] "PB(R6)"            "b'L(R)"            "dACA(R)"          
#> [121] "VMNP"              "VES(L)"            "g2(R)"            
#> [124] "ICL(L)"            "aL(L)"             "a3(R)"            
#> [127] "SMP(R)"            "lACA(R)"           "SIP(R)"           
#> [130] "PB(R9)"            "ROB(R)"            "FBl5"             
#> [133] "FBl8"              "NO(L)"             "GOR(L)"           
#> [136] "NO3(R)"            "b2(R)"             "EPA(R)"           
#> [139] "RUB(R)"            "EBr6"              "PB(L1)"           
#> [142] "CRE(L)"            "PVLP(R)"           "NO2(L)"           
#> [145] "SPS(R)"            "FLA(R)"            "PB"               
#> [148] "EBr3d"             "NO"                "AB(L)"            
#> [151] "PB(L5)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-VC1(R)"         "lACA(R)"           "g4(R)"            
#>   [4] "AL-DC4(L)"         "FBl1"              "NO(L)"            
#>   [7] "AL-DA3(R)"         "PB(R3)"            "NO2(R)"           
#>  [10] "MB(R)"             "AL-VP4(R)"         "POC"              
#>  [13] "AL-DM3(L)"         "CA(R)"             "ICL(R)"           
#>  [16] "CAN(R)"            "LAL(R)"            "AL-DA2(R)"        
#>  [19] "PB(R7)"            "PB(L8)"            "PVLP(R)"          
#>  [22] "MB(+ACA)(R)"       "SAD(-AMMC)"        "a'L(L)"           
#>  [25] "AL-DM2(R)"         "AL-VA5(R)"         "AL-VC3l(R)"       
#>  [28] "EBr1"              "AL-DC2(R)"         "AL-DM1(R)"        
#>  [31] "AL-D(R)"           "LAL(-GA)(R)"       "PB(R8)"           
#>  [34] "FBl4"              "AL-DC3(R)"         "AL-DP1m(R)"       
#>  [37] "AL-DL5(L)"         "a'1(R)"            "a3(R)"            
#>  [40] "AL-VP1m(R)"        "vACA(R)"           "SAD"              
#>  [43] "SIP(L)"            "PLP(R)"            "AL-DL2d(R)"       
#>  [46] "AL-VM2(R)"         "EBr3am"            "PB(L3)"           
#>  [49] "EBr2r4"            "AL-DL4(L)"         "g5(R)"            
#>  [52] "AME(R)"            "PB"                "NO"               
#>  [55] "FB-column3"        "aL(L)"             "CRE(R)"           
#>  [58] "CX"                "RUB(L)"            "SCL(L)"           
#>  [61] "g3(R)"             "AL-VM7d(L)"        "AL-VA7m(R)"       
#>  [64] "AL-VC4(R)"         "AL-VP2(R)"         "IB"               
#>  [67] "aL(R)"             "AL-DP1m(L)"        "AL-DP1l(R)"       
#>  [70] "ATL(L)"            "AL-VA6(L)"         "NO3(L)"           
#>  [73] "CRE(-ROB,-RUB)(R)" "PB(R6)"            "SIP(R)"           
#>  [76] "OL(R)"             "AL-DL5(R)"         "mALT(L)"          
#>  [79] "b2(R)"             "AL-VP3(R)"         "PB(R5)"           
#>  [82] "PB(L2)"            "EPA(R)"            "ROB(R)"           
#>  [85] "AVLP(R)"           "AL-DL4(R)"         "AL-VM4(R)"        
#>  [88] "AOT(R)"            "b'1(R)"            "GF(R)"            
#>  [91] "NO3(R)"            "VES(R)"            "ATL(R)"           
#>  [94] "EPA(L)"            "AL-VP1l(R)"        "LAL(L)"           
#>  [97] "AL-DM1(L)"         "AL-DA1(R)"         "AL-DM3(R)"        
#> [100] "AL-VC2(R)"         "AL-VM1(R)"         "EBr6"             
#> [103] "AL-V(R)"           "b'L(R)"            "AL-D(L)"          
#> [106] "AB(R)"             "AL-VA2(R)"         "AL-VA3(R)"        
#> [109] "AL-VL1(R)"         "SMP(L)"            "SLP(R)"           
#> [112] "FB"                "PB(L7)"            "AL-VM5v(R)"       
#> [115] "AL-VM7d(R)"        "AL-DC4(R)"         "BU(R)"            
#> [118] "AL-VC3m(R)"        "PB(L5)"            "AL-VP1d(R)"       
#> [121] "PENP"              "FBl5"              "AL(R)"            
#> [124] "AL-VL2a(R)"        "AL-DM6(L)"         "bL(L)"            
#> [127] "FBl8"              "AL-VM5d(R)"        "g1(R)"            
#> [130] "gL(L)"             "PB(L9)"            "FBl9"             
#> [133] "AL-DL3(R)"         "SNP(R)"            "a1(R)"            
#> [136] "AOTU(R)"           "LH(R)"             "LX(L)"            
#> [139] "NO2(L)"            "LX(R)"             "PB(R2)"           
#> [142] "AL-DC1(R)"         "AL-VA4(R)"         "AL-VM7v(R)"       
#> [145] "EBr3d"             "AL-DM5(L)"         "AL-VM3(R)"        
#> [148] "AL-VP5(R)"         "mALT(R)"           "GC"               
#> [151] "MB(L)"             "BU(L)"             "AL-DA2(L)"        
#> [154] "hemibrain"         "IPS(R)"            "AL-VA7l(R)"       
#> [157] "NO1(L)"            "NO1(R)"            "EB"               
#> [160] "PB(R9)"            "AL-DL1(R)"         "AL-DM4(R)"        
#> [163] "AL-VL2p(R)"        "SCL(R)"            "PB(L1)"           
#> [166] "RUB(R)"            "EBr5"              "PB(R1)"           
#> [169] "AL-DA4m(R)"        "PRW"               "INP"              
#> [172] "AL-DL2v(R)"        "dACA(R)"           "SPS(L)"           
#> [175] "b'L(L)"            "FLA(R)"            "AL-VM7v(L)"       
#> [178] "LOP(R)"            "AL-DC2(L)"         "PB(L6)"           
#> [181] "AL-VA6(R)"         "AMMC"              "PED(R)"           
#> [184] "ICL(L)"            "PB(R4)"            "AL-DM5(R)"        
#> [187] "AL(L)"             "FBl2"              "VES(L)"           
#> [190] "SNP(L)"            "FBl6"              "b1(R)"            
#> [193] "AL-DM4(L)"         "gL(R)"             "AL-DA4l(R)"       
#> [196] "AB(L)"             "AL-VA1d(R)"        "a'3(R)"           
#> [199] "CA(L)"             "VMNP"              "ME(R)"            
#> [202] "AL-DA3(L)"         "AL-DC1(L)"         "LO(R)"            
#> [205] "FBl3"              "b'2(R)"            "a'L(R)"           
#> [208] "AL-DM6(R)"         "SPS(R)"            "NO(R)"            
#> [211] "bL(R)"             "AL-VC5(R)"         "a'2(R)"           
#> [214] "AL-DA4m(L)"        "GOR(R)"            "VLNP(R)"          
#> [217] "SMP(R)"            "GOR(L)"            "g2(R)"            
#> [220] "WED(R)"            "AL-DM2(L)"         "GNG"              
#> [223] "PB(L4)"            "AL-VA1v(R)"        "a2(R)"            
#> [226] "GA(R)"             "CRE(L)"            "CRE(-RUB)(L)"     
#> [229] "EBr3pw"            "FBl7"             
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-VM7v(L)"        "FBl9"              "PB(R3)"           
#>   [4] "AL-VM2(R)"         "AL-VM5v(R)"        "b'2(R)"           
#>   [7] "mALT(R)"           "SCL(R)"            "aL(L)"            
#>  [10] "AL-D(L)"           "NO3(L)"            "FBl5"             
#>  [13] "ATL(R)"            "AL-DL2d(R)"        "RUB(R)"           
#>  [16] "PB(R8)"            "AL-DC4(L)"         "MB(+ACA)(R)"      
#>  [19] "AL-VC2(R)"         "AL-VM3(R)"         "lACA(R)"          
#>  [22] "SIP(R)"            "AL-DP1l(R)"        "FB-column3"       
#>  [25] "PB(R5)"            "EBr3am"            "bL(L)"            
#>  [28] "SCL(L)"            "AL-DM4(L)"         "AL-DM6(L)"        
#>  [31] "AL-VA6(L)"         "AL-VA4(R)"         "AL-VM4(R)"        
#>  [34] "NO"                "PB(R9)"            "LX(L)"            
#>  [37] "NO2(L)"            "NO2(R)"            "AL-DL3(R)"        
#>  [40] "a'3(R)"            "AL-VA2(R)"         "EB"               
#>  [43] "ICL(R)"            "AL-VA1d(R)"        "AL-VP1d(R)"       
#>  [46] "mALT(L)"           "FLA(R)"            "AL-DP1m(L)"       
#>  [49] "NO(L)"             "LAL(-GA)(R)"       "VMNP"             
#>  [52] "AVLP(R)"           "AL-D(R)"           "AL-VM7d(R)"       
#>  [55] "b'L(L)"            "AL-DM1(R)"         "CRE(-ROB,-RUB)(R)"
#>  [58] "PRW"               "ME(R)"             "AL-DL5(L)"        
#>  [61] "FBl3"              "FBl4"              "GNG"              
#>  [64] "AL-DP1m(R)"        "AL-VC3l(R)"        "NO1(L)"           
#>  [67] "AL-DM3(L)"         "AL-DM6(R)"         "AL-V(R)"          
#>  [70] "AB(L)"             "BU(R)"             "AL-DL4(L)"        
#>  [73] "AL-DM1(L)"         "AL-DA2(L)"         "PB(R2)"           
#>  [76] "a1(R)"             "IPS(R)"            "AL-DL1(R)"        
#>  [79] "AL-VM5d(R)"        "IB"                "g3(R)"            
#>  [82] "gL(L)"             "AL-DL4(R)"         "PB(L3)"           
#>  [85] "AL-DM2(R)"         "AL-VL2a(R)"        "VLNP(R)"          
#>  [88] "LX(R)"             "EBr3d"             "EPA(R)"           
#>  [91] "a'L(L)"            "bL(R)"             "AL-DC3(R)"        
#>  [94] "g1(R)"             "AL-DC1(L)"         "NO(R)"            
#>  [97] "a'L(R)"            "AL-VP4(R)"         "PB(R1)"           
#> [100] "a3(R)"             "AL-VC4(R)"         "PENP"             
#> [103] "SIP(L)"            "AL-DM4(R)"         "AOTU(R)"          
#> [106] "AME(R)"            "CRE(L)"            "AL(R)"            
#> [109] "AMMC"              "SPS(L)"            "g4(R)"            
#> [112] "PB"                "PB(L7)"            "MB(L)"            
#> [115] "EBr2r4"            "AL-VC5(R)"         "SAD"              
#> [118] "aL(R)"             "CRE(R)"            "AL-DC2(R)"        
#> [121] "AL-DM3(R)"         "EBr3pw"            "LO(R)"            
#> [124] "AL-DA3(R)"         "AL-VM1(R)"         "AL-DA3(L)"        
#> [127] "gL(R)"             "GOR(R)"            "LH(R)"            
#> [130] "SMP(L)"            "EBr1"              "ICL(L)"           
#> [133] "SNP(R)"            "AL-VC3m(R)"        "AL-VL2p(R)"       
#> [136] "POC"               "SMP(R)"            "b'1(R)"           
#> [139] "GOR(L)"            "ROB(R)"            "AL-DA4l(R)"       
#> [142] "FB"                "BU(L)"             "PB(L1)"           
#> [145] "AOT(R)"            "SNP(L)"            "MB(R)"            
#> [148] "AL-VP3(R)"         "NO3(R)"            "b2(R)"            
#> [151] "AL-VP1l(R)"        "AL-DA2(R)"         "AL-DA4m(R)"       
#> [154] "a2(R)"             "FBl1"              "ATL(L)"           
#> [157] "EBr5"              "a'2(R)"            "AL-DL2v(R)"       
#> [160] "PB(R4)"            "AL-VA7l(R)"        "AL-VL1(R)"        
#> [163] "CA(L)"             "PB(R6)"            "b'L(R)"           
#> [166] "PB(L6)"            "FBl2"              "AL-VA1v(R)"       
#> [169] "vACA(R)"           "PLP(R)"            "CRE(-RUB)(L)"     
#> [172] "g5(R)"             "FBl6"              "FBl7"             
#> [175] "CA(R)"             "LOP(R)"            "GF(R)"            
#> [178] "AL(L)"             "PB(L9)"            "dACA(R)"          
#> [181] "AL-VP1m(R)"        "CX"                "AL-VA7m(R)"       
#> [184] "WED(R)"            "AL-VA3(R)"         "AL-VA6(R)"        
#> [187] "SLP(R)"            "LAL(L)"            "EBr6"             
#> [190] "b1(R)"             "hemibrain"         "VES(R)"           
#> [193] "EPA(L)"            "AL-VA5(R)"         "AL-DM2(L)"        
#> [196] "LAL(R)"            "NO1(R)"            "AL-VC1(R)"        
#> [199] "AL-VP2(R)"         "AL-DC1(R)"         "AL-DL5(R)"        
#> [202] "AL-DM5(L)"         "FBl8"              "AL-DA1(R)"        
#> [205] "GA(R)"             "PB(L2)"            "INP"              
#> [208] "AL-VM7v(R)"        "AL-DA4m(L)"        "AL-VM7d(L)"       
#> [211] "CAN(R)"            "SPS(R)"            "g2(R)"            
#> [214] "AB(R)"             "AL-VP5(R)"         "OL(R)"            
#> [217] "VES(L)"            "AL-DM5(R)"         "PB(R7)"           
#> [220] "GC"                "AL-DC2(L)"         "PB(L5)"           
#> [223] "PED(R)"            "SAD(-AMMC)"        "AL-DC4(R)"        
#> [226] "PB(L8)"            "PB(L4)"            "a'1(R)"           
#> [229] "PVLP(R)"           "RUB(L)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method   path                                        
#> [1,] "GET"    "/api"                                      
#> [2,] "REPORT" "/api/*"                                    
#> [3,] "GET"    "/api/v:ver/raw/keyvalue/key/:instance/:key"
#> [4,] "POST"   "/api/v:ver/npexplorer/findneurons"         
#> [5,] "POST"   "/api/v:ver/raw/cypher/transaction"         
#> [6,] "GET"    "/api/*"                                    
#>      name                                                                
#> [1,] "github.com/labstack/echo.glob..func1"                              
#> [2,] "github.com/labstack/echo.glob..func1"                              
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [6,] "github.com/labstack/echo.glob..func1"                              

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }