Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0`
#> $`hemibrain:v1.0`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0`$ROIs
#>   [1] "SLP(R)"            "PB(L2)"            "SPS(R)"           
#>   [4] "CA(R)"             "g5(R)"             "OL(R)"            
#>   [7] "vACA(R)"           "BU(L)"             "PLP(R)"           
#>  [10] "PB(L9)"            "FBl3"              "FBl8"             
#>  [13] "LAL(R)"            "AOT(R)"            "AB(L)"            
#>  [16] "EBr3d"             "a'L(R)"            "ATL(R)"           
#>  [19] "aL(L)"             "LOP(R)"            "LAL(-GA)(R)"      
#>  [22] "SAD(-AMMC)"        "NO3(L)"            "NO2(L)"           
#>  [25] "lACA(R)"           "mALT(R)"           "ICL(L)"           
#>  [28] "PB(L7)"            "EBr3am"            "PENP"             
#>  [31] "SIP(R)"            "AB(R)"             "EB"               
#>  [34] "MB(+ACA)(R)"       "INP"               "bL(R)"            
#>  [37] "AMMC"              "CA(L)"             "b'2(R)"           
#>  [40] "NO1(L)"            "CRE(L)"            "IPS(R)"           
#>  [43] "NO2(R)"            "a'3(R)"            "PB(L3)"           
#>  [46] "PB(L6)"            "FB"                "EBr2r4"           
#>  [49] "WED(R)"            "NO3(R)"            "GA(R)"            
#>  [52] "SNP(L)"            "AME(R)"            "SCL(L)"           
#>  [55] "SMP(L)"            "a2(R)"             "BU(R)"            
#>  [58] "VES(R)"            "a1(R)"             "PED(R)"           
#>  [61] "a'2(R)"            "CX"                "NO"               
#>  [64] "FBl5"              "b2(R)"             "b'L(R)"           
#>  [67] "b1(R)"             "PB(R5)"            "SIP(L)"           
#>  [70] "LO(R)"             "ME(R)"             "AVLP(R)"          
#>  [73] "g3(R)"             "PB(L1)"            "b'L(L)"           
#>  [76] "PB(L4)"            "SMP(R)"            "ATL(L)"           
#>  [79] "EBr1"              "NO(R)"             "bL(L)"            
#>  [82] "FBl2"              "FBl9"              "VLNP(R)"          
#>  [85] "LAL(L)"            "PRW"               "FBl1"             
#>  [88] "a3(R)"             "NO(L)"             "VES(L)"           
#>  [91] "EPA(R)"            "gL(R)"             "LH(R)"            
#>  [94] "LX(L)"             "AL(R)"             "POC"              
#>  [97] "b'1(R)"            "SNP(R)"            "AL(L)"            
#> [100] "dACA(R)"           "PB(R9)"            "PB(R1)"           
#> [103] "EPA(L)"            "GF(R)"             "hemibrain"        
#> [106] "MB(R)"             "CRE(-ROB,-RUB)(R)" "EBr6"             
#> [109] "PB(R2)"            "PB"                "VMNP"             
#> [112] "a'L(L)"            "IB"                "PB(L5)"           
#> [115] "GC"                "MB(L)"             "RUB(R)"           
#> [118] "EBr5"              "CAN(R)"            "RUB(L)"           
#> [121] "PB(R4)"            "SPS(L)"            "FBl7"             
#> [124] "gL(L)"             "GNG"               "aL(R)"            
#> [127] "SCL(R)"            "LX(R)"             "g4(R)"            
#> [130] "EBr3pw"            "GOR(L)"            "PB(R8)"           
#> [133] "ICL(R)"            "PB(R7)"            "NO1(R)"           
#> [136] "a'1(R)"            "PVLP(R)"           "mALT(L)"          
#> [139] "AOTU(R)"           "g2(R)"             "ROB(R)"           
#> [142] "PB(R3)"            "GOR(R)"            "g1(R)"            
#> [145] "PB(L8)"            "FBl6"              "CRE(R)"           
#> [148] "FBl4"              "FLA(R)"            "PB(R6)"           
#> [151] "SAD"              
#> 
#> $`hemibrain:v1.0`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                                  
#> [1,] "POST"    "/api/npexplorer/distribution"        
#> [2,] "GET"     "/token"                              
#> [3,] "CONNECT" "/api/*"                              
#> [4,] "GET"     "/api/v:ver/cached/dailytype"         
#> [5,] "POST"    "/api/npexplorer/commonconnectivity"  
#> [6,] "GET"     "/api/skeletons/skeleton/:dataset/:id"
#>      name                                                                                              
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"      
#> [2,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                          
#> [3,] "github.com/labstack/echo.glob..func1"                                                            
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCommonConnectivity-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/skeletons.masterAPI.getSkeleton-fm"              

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }