Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0`
#> $`hemibrain:v1.0`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0`$ROIs
#>   [1] "EBr2r4"            "EBr5"              "a'1(R)"           
#>   [4] "IB"                "SIP(L)"            "g3(R)"            
#>   [7] "g2(R)"             "PB(L8)"            "SMP(L)"           
#>  [10] "EBr1"              "g4(R)"             "PB(R1)"           
#>  [13] "PB(R4)"            "EBr3am"            "LAL(-GA)(R)"      
#>  [16] "EBr6"              "PB(R9)"            "ICL(L)"           
#>  [19] "CX"                "GF(R)"             "EBr3pw"           
#>  [22] "LAL(R)"            "MB(L)"             "CRE(R)"           
#>  [25] "b'L(R)"            "EPA(R)"            "CA(R)"            
#>  [28] "PB(R2)"            "LOP(R)"            "aL(R)"            
#>  [31] "GC"                "a'L(R)"            "ATL(L)"           
#>  [34] "lACA(R)"           "SNP(R)"            "FBl5"             
#>  [37] "b'L(L)"            "NO(L)"             "AL(R)"            
#>  [40] "SCL(L)"            "SAD"               "FBl6"             
#>  [43] "VES(R)"            "FBl1"              "AME(R)"           
#>  [46] "LX(L)"             "PB(R5)"            "CA(L)"            
#>  [49] "g1(R)"             "SPS(L)"            "PB"               
#>  [52] "PVLP(R)"           "RUB(L)"            "a1(R)"            
#>  [55] "GA(R)"             "POC"               "OL(R)"            
#>  [58] "PRW"               "SCL(R)"            "SAD(-AMMC)"       
#>  [61] "GNG"               "b'1(R)"            "b'2(R)"           
#>  [64] "NO1(L)"            "LH(R)"             "ATL(R)"           
#>  [67] "VES(L)"            "SMP(R)"            "AOTU(R)"          
#>  [70] "PB(R7)"            "ME(R)"             "g5(R)"            
#>  [73] "SLP(R)"            "PB(L5)"            "hemibrain"        
#>  [76] "PED(R)"            "LX(R)"             "AVLP(R)"          
#>  [79] "RUB(R)"            "PB(R6)"            "PB(L2)"           
#>  [82] "aL(L)"             "a'L(L)"            "FBl8"             
#>  [85] "PB(L1)"            "gL(L)"             "SIP(R)"           
#>  [88] "PB(L9)"            "PB(L7)"            "FBl3"             
#>  [91] "IPS(R)"            "LAL(L)"            "GOR(L)"           
#>  [94] "WED(R)"            "EBr3d"             "AL(L)"            
#>  [97] "NO"                "LO(R)"             "FBl2"             
#> [100] "a2(R)"             "NO1(R)"            "BU(R)"            
#> [103] "ROB(R)"            "PB(L6)"            "ICL(R)"           
#> [106] "PENP"              "a'2(R)"            "SNP(L)"           
#> [109] "FBl7"              "NO2(L)"            "b2(R)"            
#> [112] "AB(R)"             "bL(R)"             "PB(L4)"           
#> [115] "gL(R)"             "VMNP"              "NO3(L)"           
#> [118] "NO2(R)"            "GOR(R)"            "AOT(R)"           
#> [121] "PLP(R)"            "vACA(R)"           "CRE(-ROB,-RUB)(R)"
#> [124] "AB(L)"             "FLA(R)"            "FBl9"             
#> [127] "VLNP(R)"           "AMMC"              "PB(L3)"           
#> [130] "CRE(L)"            "MB(+ACA)(R)"       "b1(R)"            
#> [133] "NO3(R)"            "dACA(R)"           "bL(L)"            
#> [136] "CAN(R)"            "a'3(R)"            "mALT(R)"          
#> [139] "PB(R8)"            "INP"               "SPS(R)"           
#> [142] "EB"                "FBl4"              "PB(R3)"           
#> [145] "EPA(L)"            "MB(R)"             "FB"               
#> [148] "NO(R)"             "a3(R)"             "mALT(L)"          
#> [151] "BU(L)"            
#> 
#> $`hemibrain:v1.0`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                
#> [1,] "GET"  "/api"                              
#> [2,] "GET"  "/api/v:ver/dbmeta/version"         
#> [3,] "GET"  "/api/dbmeta/datasets"              
#> [4,] "GET"  "/api/v:ver/npexplorer/neuronmeta"  
#> [5,] "GET"  "/api/v:ver/npexplorer/roisinneuron"
#> [6,] "GET"  "/api/npexplorer/distribution"      
#>      name                                                                                        
#> [1,] "github.com/labstack/echo.glob..func1"                                                      
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatasets-fm"            
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }