Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "NO"                "g5(R)"             "SNP(L)"           
#>   [4] "ATL(R)"            "b1(R)"             "PB(R6)"           
#>   [7] "PB(L3)"            "PB(R3)"            "EPA(L)"           
#>  [10] "PB(L4)"            "EBr6"              "PB(R1)"           
#>  [13] "PB(L1)"            "PB(R5)"            "LAL(L)"           
#>  [16] "MB(L)"             "FBl7"              "IB"               
#>  [19] "SMP(L)"            "PB(R8)"            "VMNP"             
#>  [22] "a3(R)"             "LAL(R)"            "AL(R)"            
#>  [25] "g3(R)"             "NO3(R)"            "bL(L)"            
#>  [28] "PVLP(R)"           "WED(R)"            "BU(R)"            
#>  [31] "MB(R)"             "ROB(R)"            "IPS(R)"           
#>  [34] "SPS(L)"            "VES(L)"            "PENP"             
#>  [37] "FLA(R)"            "FBl9"              "aL(L)"            
#>  [40] "POC"               "AOT(R)"            "AB(L)"            
#>  [43] "a1(R)"             "SPS(R)"            "LAL(-GA)(R)"      
#>  [46] "aL(R)"             "EBr3pw"            "LOP(R)"           
#>  [49] "SIP(L)"            "PB(R7)"            "ME(R)"            
#>  [52] "EBr2r4"            "a'1(R)"            "CA(R)"            
#>  [55] "GF(R)"             "GA(R)"             "GNG"              
#>  [58] "AME(R)"            "mALT(L)"           "CA(L)"            
#>  [61] "NO2(L)"            "b2(R)"             "FBl3"             
#>  [64] "FB"                "VES(R)"            "FBl1"             
#>  [67] "INP"               "RUB(R)"            "EBr5"             
#>  [70] "PB(R9)"            "PB(L2)"            "SNP(R)"           
#>  [73] "PB(L6)"            "LX(L)"             "b'L(L)"           
#>  [76] "EBr3am"            "ATL(L)"            "gL(R)"            
#>  [79] "PB(R2)"            "hemibrain"         "g4(R)"            
#>  [82] "NO1(R)"            "ICL(L)"            "AB(R)"            
#>  [85] "NO2(R)"            "SMP(R)"            "vACA(R)"          
#>  [88] "AMMC"              "g2(R)"             "CRE(L)"           
#>  [91] "CRE(-ROB,-RUB)(R)" "PB(L9)"            "LO(R)"            
#>  [94] "a'L(L)"            "CX"                "AOTU(R)"          
#>  [97] "FBl2"              "GC"                "NO1(L)"           
#> [100] "CRE(R)"            "a'L(R)"            "NO(R)"            
#> [103] "SLP(R)"            "b'2(R)"            "SAD"              
#> [106] "SAD(-AMMC)"        "MB(+ACA)(R)"       "FBl6"             
#> [109] "AVLP(R)"           "ICL(R)"            "a2(R)"            
#> [112] "NO3(L)"            "b'L(R)"            "PLP(R)"           
#> [115] "GOR(L)"            "EBr1"              "LX(R)"            
#> [118] "EPA(R)"            "AL(L)"             "PRW"              
#> [121] "PED(R)"            "dACA(R)"           "PB(L8)"           
#> [124] "gL(L)"             "a'3(R)"            "lACA(R)"          
#> [127] "BU(L)"             "FBl4"              "FBl8"             
#> [130] "CAN(R)"            "PB(L5)"            "mALT(R)"          
#> [133] "PB(L7)"            "LH(R)"             "SCL(R)"           
#> [136] "a'2(R)"            "PB(R4)"            "OL(R)"            
#> [139] "VLNP(R)"           "PB"                "RUB(L)"           
#> [142] "EB"                "FBl5"              "bL(R)"            
#> [145] "GOR(R)"            "NO(L)"             "b'1(R)"           
#> [148] "g1(R)"             "EBr3d"             "SIP(R)"           
#> [151] "SCL(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                   
#> [1,] "GET"  "/api/npexplorer/rankedtable"          
#> [2,] "POST" "/api/v:ver/npexplorer/roiconnectivity"
#> [3,] "GET"  "/api/npexplorer/autapses"             
#> [4,] "GET"  "/api/v:ver/dbmeta/database"           
#> [5,] "GET"  "/api/v:ver/npexplorer/neuronmeta"     
#> [6,] "GET"  "/api/npexplorer/commonconnectivity"   
#>      name                                                                                              
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getRankedTable-fm"       
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"          
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCommonConnectivity-fm"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }