Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "FBl8"              "AL-VM4(R)"         "PB(R5)"           
#>   [4] "PB(L2)"            "GC"                "VMNP"             
#>   [7] "PB(L9)"            "VLNP(R)"           "AL-DA2(L)"        
#>  [10] "AL-VM7d(R)"        "CA(L)"             "EBr3d"            
#>  [13] "VES(L)"            "AL-DL4(R)"         "CX"               
#>  [16] "lACA(R)"           "GOR(L)"            "AL-DM1(L)"        
#>  [19] "AL-DC2(R)"         "AL-VA5(R)"         "AL-VC3l(R)"       
#>  [22] "NO2(L)"            "b'L(R)"            "ROB(R)"           
#>  [25] "PB(R1)"            "AL-DL3(R)"         "AL-VA1d(R)"       
#>  [28] "PB(L3)"            "FLA(R)"            "AL-DL4(L)"        
#>  [31] "AL-VA4(R)"         "PENP"              "PLP(R)"           
#>  [34] "AL(R)"             "AL-DA4l(R)"        "NO(R)"            
#>  [37] "CRE(-RUB)(L)"      "AL-VM7v(R)"        "AL-VP5(R)"        
#>  [40] "a2(R)"             "MB(L)"             "EB"               
#>  [43] "NO3(L)"            "FBl7"              "AVLP(R)"          
#>  [46] "EPA(L)"            "AL-DC3(R)"         "AL-DL5(R)"        
#>  [49] "AL-VC1(R)"         "AL-VM5v(R)"        "AL-DM5(R)"        
#>  [52] "AB(L)"             "vACA(R)"           "LAL(L)"           
#>  [55] "AL-VA6(L)"         "AL-DA4m(R)"        "AL-DL2d(R)"       
#>  [58] "AL-VP1l(R)"        "mALT(L)"           "NO3(R)"           
#>  [61] "AL-V(R)"           "AL-DL5(L)"         "AL-DM6(L)"        
#>  [64] "g5(R)"             "PB(L4)"            "b2(R)"            
#>  [67] "bL(L)"             "b'L(L)"            "PB(L8)"           
#>  [70] "AL-DM6(R)"         "AL-VA6(R)"         "b'1(R)"           
#>  [73] "a'1(R)"            "LH(R)"             "NO(L)"            
#>  [76] "gL(L)"             "AL-DC4(L)"         "g4(R)"            
#>  [79] "CRE(L)"            "SLP(R)"            "BU(L)"            
#>  [82] "ME(R)"             "ATL(R)"            "BU(R)"            
#>  [85] "AL-VM7v(L)"        "AME(R)"            "PB"               
#>  [88] "PVLP(R)"           "b1(R)"             "AL-VC3m(R)"       
#>  [91] "AL-VL2p(R)"        "SPS(R)"            "AL-D(L)"          
#>  [94] "AL-DP1m(L)"        "NO2(R)"            "SCL(L)"           
#>  [97] "hemibrain"         "AL-DC2(L)"         "AL-DA1(R)"        
#> [100] "AL-DA3(R)"         "AL-VM5d(R)"        "AL-VP4(R)"        
#> [103] "PB(L5)"            "CRE(-ROB,-RUB)(R)" "SAD"              
#> [106] "IB112"             "AL-DM5(L)"         "AL-VM1(R)"        
#> [109] "PB(R6)"            "aL(R)"             "CRE(R)"           
#> [112] "FB"                "SAD(-AMMC)"        "FBl9"             
#> [115] "AL-VA7m(R)"        "AL-VP2(R)"         "CAN(R)"           
#> [118] "PB(L1)"            "AL-VA2(R)"         "EBr6"             
#> [121] "SNP(R)"            "AL-VC4(R)"         "SMP(R)"           
#> [124] "AOTU(R)"           "AL-DM2(R)"         "LOP(R)"           
#> [127] "a1(R)"             "LO(R)"             "MB(R)"            
#> [130] "AL-VC5(R)"         "b'2(R)"            "PB(R7)"           
#> [133] "MB(+ACA)(R)"       "aL(L)"             "SCL(R)"           
#> [136] "LX(R)"             "PB(R3)"            "a'L(L)"           
#> [139] "AL-DC1(R)"         "AL-VA3(R)"         "gL(R)"            
#> [142] "AL-DM3(R)"         "AL-VP1m(R)"        "AMMC"             
#> [145] "LAL(-GA)(R)"       "a'2(R)"            "AL-DA3(L)"        
#> [148] "PB(L7)"            "AL-VA7l(R)"        "mALT(R)"          
#> [151] "AOT(R)"            "RUB(R)"            "EBr5"             
#> [154] "FBl2"              "AL-VM2(R)"         "SIP(L)"           
#> [157] "g2(R)"             "a'L(R)"            "AL-DM3(L)"        
#> [160] "AB(R)"             "IPS(R)"            "AL-DM1(R)"        
#> [163] "GF(R)"             "EPA(R)"            "AL-DA4m(L)"       
#> [166] "ICL(R)"            "bL(R)"             "OL(R)"            
#> [169] "AL-VL1(R)"         "EBr3am"            "PED(R)"           
#> [172] "PB(L6)"            "GA(R)"             "PRW"              
#> [175] "EBr2r4"            "AL-DM4(L)"         "PB(R9)"           
#> [178] "EBr1"              "CA(R)"             "GOR(R)"           
#> [181] "AL-DA2(R)"         "VES(R)"            "AL-DM4(R)"        
#> [184] "AL-VC2(R)"         "AL-VM3(R)"         "SIP(R)"           
#> [187] "EBr3pw"            "ICL(L)"            "AL-DP1l(R)"       
#> [190] "AL-VL2a(R)"        "a'3(R)"            "NO1(L)"           
#> [193] "NO"                "INP"               "ATL(L)"           
#> [196] "SPS(L)"            "FBl3"              "IB"               
#> [199] "WED(R)"            "AL-VA1v(R)"        "NO1(R)"           
#> [202] "FBl5"              "PB(R2)"            "POC"              
#> [205] "SMP(L)"            "AL-DP1m(R)"        "FB-column3"       
#> [208] "FBl1"              "RUB(L)"            "AL-DL1(R)"        
#> [211] "FBl4"              "g1(R)"             "dACA(R)"          
#> [214] "AL-VM7d(L)"        "AL-DC4(R)"         "AL-D(R)"          
#> [217] "AL-VP1d(R)"        "GNG"               "LAL(R)"           
#> [220] "AL-VP3(R)"         "FBl6"              "PB(R4)"           
#> [223] "a3(R)"             "PB(R8)"            "AL-DC1(L)"        
#> [226] "AL-DM2(L)"         "AL-DL2v(R)"        "g3(R)"            
#> [229] "AL(L)"             "SNP(L)"            "LX(L)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "FBl9"              "LH(R)"             "SPS(R)"           
#>   [4] "g1(R)"             "MB(L)"             "FB"               
#>   [7] "AL(L)"             "NO3(L)"            "b1(R)"            
#>  [10] "a3(R)"             "POC"               "PB(R7)"           
#>  [13] "g5(R)"             "FBl5"              "FBl6"             
#>  [16] "AB(R)"             "EPA(L)"            "GC"               
#>  [19] "SCL(R)"            "FBl1"              "VES(L)"           
#>  [22] "PB(L1)"            "PB"                "NO"               
#>  [25] "PVLP(R)"           "PB(R3)"            "NO2(L)"           
#>  [28] "AOTU(R)"           "NO1(L)"            "BU(L)"            
#>  [31] "FBl4"              "OL(R)"             "mALT(L)"          
#>  [34] "PB(R6)"            "PB(L3)"            "g4(R)"            
#>  [37] "SNP(L)"            "PB(R2)"            "AMMC"             
#>  [40] "CX"                "b'2(R)"            "SAD"              
#>  [43] "PB(L8)"            "SPS(L)"            "RUB(L)"           
#>  [46] "IB"                "LAL(L)"            "EBr1"             
#>  [49] "SMP(R)"            "gL(L)"             "PB(L9)"           
#>  [52] "CA(R)"             "CRE(-ROB,-RUB)(R)" "a1(R)"            
#>  [55] "EBr6"              "EBr3d"             "PB(R1)"           
#>  [58] "a'1(R)"            "b'L(L)"            "SLP(R)"           
#>  [61] "CA(L)"             "GA(R)"             "PB(R8)"           
#>  [64] "LX(R)"             "FBl2"              "FBl7"             
#>  [67] "g2(R)"             "VMNP"              "PB(R9)"           
#>  [70] "NO(R)"             "bL(L)"             "FBl3"             
#>  [73] "LAL(-GA)(R)"       "aL(L)"             "LOP(R)"           
#>  [76] "vACA(R)"           "PRW"               "a'2(R)"           
#>  [79] "CAN(R)"            "a'L(L)"            "mALT(R)"          
#>  [82] "aL(R)"             "b'L(R)"            "PLP(R)"           
#>  [85] "AVLP(R)"           "a'3(R)"            "bL(R)"            
#>  [88] "GF(R)"             "lACA(R)"           "EBr5"             
#>  [91] "dACA(R)"           "EB"                "SIP(R)"           
#>  [94] "PB(L4)"            "MB(R)"             "PED(R)"           
#>  [97] "EBr2r4"            "PB(L6)"            "VES(R)"           
#> [100] "ATL(R)"            "LAL(R)"            "g3(R)"            
#> [103] "GOR(L)"            "AOT(R)"            "b'1(R)"           
#> [106] "PB(L2)"            "WED(R)"            "MB(+ACA)(R)"      
#> [109] "hemibrain"         "ICL(L)"            "ICL(R)"           
#> [112] "PB(R4)"            "CRE(L)"            "LO(R)"            
#> [115] "NO1(R)"            "IPS(R)"            "INP"              
#> [118] "GNG"               "PENP"              "ROB(R)"           
#> [121] "gL(R)"             "AL(R)"             "PB(R5)"           
#> [124] "BU(R)"             "AME(R)"            "RUB(R)"           
#> [127] "EBr3am"            "NO3(R)"            "EPA(R)"           
#> [130] "PB(L7)"            "SAD(-AMMC)"        "FBl8"             
#> [133] "VLNP(R)"           "CRE(R)"            "b2(R)"            
#> [136] "ME(R)"             "NO2(R)"            "GOR(R)"           
#> [139] "PB(L5)"            "SNP(R)"            "SCL(L)"           
#> [142] "a2(R)"             "SIP(L)"            "LX(L)"            
#> [145] "AB(L)"             "SMP(L)"            "FLA(R)"           
#> [148] "EBr3pw"            "a'L(R)"            "NO(L)"            
#> [151] "ATL(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "EBr3pw"            "AL-VC3l(R)"        "b'2(R)"           
#>   [4] "lACA(R)"           "NO1(R)"            "AL-DM3(R)"        
#>   [7] "AL-VP1m(R)"        "LAL(L)"            "AL-DC2(L)"        
#>  [10] "AL-VM7d(L)"        "AL-DA1(R)"         "SCL(L)"           
#>  [13] "vACA(R)"           "AL-DA3(R)"         "AL-VA3(R)"        
#>  [16] "SLP(R)"            "AL-VA1v(R)"        "SPS(R)"           
#>  [19] "AL-DL5(L)"         "NO(R)"             "SIP(L)"           
#>  [22] "AL-DA4m(R)"        "SMP(R)"            "hemibrain"        
#>  [25] "AL-DA3(L)"         "AB(R)"             "PB(R1)"           
#>  [28] "GNG"               "AL-DL1(R)"         "AL-VM7d(R)"       
#>  [31] "GA(R)"             "g4(R)"             "AL-DL4(L)"        
#>  [34] "AL-VM1(R)"         "AL-VP3(R)"         "PB(R7)"           
#>  [37] "PB(R8)"            "PB(L7)"            "AL(L)"            
#>  [40] "PB(R5)"            "EBr1"              "BU(L)"            
#>  [43] "AL-VC3m(R)"        "PB(L2)"            "gL(L)"            
#>  [46] "FBl7"              "AL-V(R)"           "NO"               
#>  [49] "FLA(R)"            "VMNP"              "a'2(R)"           
#>  [52] "PB(L5)"            "AL-DA4m(L)"        "FBl8"             
#>  [55] "AL-VM7v(R)"        "EPA(L)"            "AL-DA2(L)"        
#>  [58] "AL-DP1m(L)"        "b1(R)"             "AL-D(R)"          
#>  [61] "AL-VM7v(L)"        "aL(L)"             "b'L(L)"           
#>  [64] "EBr3d"             "AL-VM2(R)"         "PVLP(R)"          
#>  [67] "PB(R4)"            "VES(L)"            "POC"              
#>  [70] "AOTU(R)"           "aL(R)"             "PB(L6)"           
#>  [73] "AL-DM4(L)"         "AL-VM4(R)"         "AL-VA2(R)"        
#>  [76] "a'3(R)"            "AL-VC4(R)"         "AL-VL1(R)"        
#>  [79] "AL-VP4(R)"         "PB(L1)"            "a'1(R)"           
#>  [82] "bL(R)"             "AL-DM4(R)"         "RUB(R)"           
#>  [85] "g2(R)"             "SCL(R)"            "EBr2r4"           
#>  [88] "AL-DM6(L)"         "SIP(R)"            "AL-DA4l(R)"       
#>  [91] "AL-VA7l(R)"        "PENP"              "b2(R)"            
#>  [94] "PB(L8)"            "MB(+ACA)(R)"       "MB(R)"            
#>  [97] "LO(R)"             "FBl6"              "a'L(R)"           
#> [100] "EPA(R)"            "bL(L)"             "FB"               
#> [103] "EBr6"              "LX(L)"             "VLNP(R)"          
#> [106] "AMMC"              "NO3(R)"            "NO1(L)"           
#> [109] "MB(L)"             "dACA(R)"           "AL-DC2(R)"        
#> [112] "IB"                "FBl2"              "NO(L)"            
#> [115] "CRE(-RUB)(L)"      "AL-VC1(R)"         "AOT(R)"           
#> [118] "PRW"               "AL-VL2a(R)"        "CRE(-ROB,-RUB)(R)"
#> [121] "NO3(L)"            "FBl1"              "INP"              
#> [124] "AL-DC4(R)"         "AL-DL4(R)"         "AL-VA7m(R)"       
#> [127] "a1(R)"             "g1(R)"             "CRE(R)"           
#> [130] "a3(R)"             "ROB(R)"            "AL-DL3(R)"        
#> [133] "a2(R)"             "ME(R)"             "BU(R)"            
#> [136] "AL-DP1l(R)"        "LX(R)"             "AL-DC1(R)"        
#> [139] "AL-DL2v(R)"        "RUB(L)"            "GOR(R)"           
#> [142] "AL-DM6(R)"         "AL-VP1l(R)"        "PB(R6)"           
#> [145] "AL-DC4(L)"         "AL-VA1d(R)"        "mALT(R)"          
#> [148] "WED(R)"            "SAD(-AMMC)"        "AL-DC1(L)"        
#> [151] "AL-VP2(R)"         "LAL(-GA)(R)"       "AL-DP1m(R)"       
#> [154] "AL-VA6(R)"         "AL-VM3(R)"         "CA(L)"            
#> [157] "PB(L4)"            "AL-DA2(R)"         "AL-DL2d(R)"       
#> [160] "AL-DM1(L)"         "ATL(R)"            "AL-VA4(R)"        
#> [163] "g3(R)"             "SNP(L)"            "AME(R)"           
#> [166] "AL-DM3(L)"         "PB(R9)"            "AVLP(R)"          
#> [169] "OL(R)"             "EBr3am"            "mALT(L)"          
#> [172] "PB"                "PB(L9)"            "g5(R)"            
#> [175] "NO2(R)"            "AL-DL5(R)"         "AL(R)"            
#> [178] "AL-VM5v(R)"        "LAL(R)"            "b'1(R)"           
#> [181] "PLP(R)"            "AL-DM5(L)"         "FBl5"             
#> [184] "PB(R3)"            "AL-DM5(R)"         "ATL(L)"           
#> [187] "FBl4"              "AL-DM1(R)"         "AL-VL2p(R)"       
#> [190] "SNP(R)"            "LH(R)"             "FB-column3"       
#> [193] "AL-VP1d(R)"        "GF(R)"             "PED(R)"           
#> [196] "SPS(L)"            "AL-D(L)"           "VES(R)"           
#> [199] "CA(R)"             "GC"                "b'L(R)"           
#> [202] "EBr5"              "ICL(R)"            "SMP(L)"           
#> [205] "SAD"               "CRE(L)"            "AL-DM2(R)"        
#> [208] "AL-VC5(R)"         "AB(L)"             "gL(R)"            
#> [211] "LOP(R)"            "AL-VM5d(R)"        "AL-DM2(L)"        
#> [214] "ICL(L)"            "a'L(L)"            "AL-VP5(R)"        
#> [217] "PB(L3)"            "IPS(R)"            "AL-VA5(R)"        
#> [220] "AL-DC3(R)"         "CX"                "NO2(L)"           
#> [223] "EB"                "AL-VA6(L)"         "FBl3"             
#> [226] "AL-VC2(R)"         "FBl9"              "CAN(R)"           
#> [229] "GOR(L)"            "PB(R2)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "SIP(L)"            "EBr5"              "AVLP(R)"          
#>   [4] "AL-DC1(R)"         "NO1(L)"            "FB-column3"       
#>   [7] "g1(R)"             "AL-DA4l(R)"        "CRE(R)"           
#>  [10] "FBl2"              "b'L(R)"            "gL(L)"            
#>  [13] "b'L(L)"            "POC"               "PRW"              
#>  [16] "AL-VM7v(L)"        "IPS(R)"            "AL-VM5d(R)"       
#>  [19] "AMMC"              "LAL(L)"            "GF(R)"            
#>  [22] "a'1(R)"            "AL-DM1(R)"         "EB"               
#>  [25] "g4(R)"             "b1(R)"             "AL-VA4(R)"        
#>  [28] "AL-VP4(R)"         "AL-DC1(L)"         "PB(L9)"           
#>  [31] "bL(L)"             "CA(R)"             "AL-DM5(R)"        
#>  [34] "AL-VC3m(R)"        "AOT(R)"            "hemibrain"        
#>  [37] "gL(R)"             "AL-DC4(R)"         "AL-VA7l(R)"       
#>  [40] "AL(R)"             "AL-VP2(R)"         "PB(L1)"           
#>  [43] "AL-DL5(L)"         "ICL(R)"            "PB(L4)"           
#>  [46] "PB(L8)"            "AL-DC2(L)"         "AL-VC3l(R)"       
#>  [49] "vACA(R)"           "AL-VM7v(R)"        "AB(L)"            
#>  [52] "PB(L3)"            "AL-DP1m(L)"        "NO3(L)"           
#>  [55] "AL-DL2d(R)"        "SCL(L)"            "ME(R)"            
#>  [58] "BU(R)"             "FBl8"              "AL-VA7m(R)"       
#>  [61] "GOR(L)"            "SLP(R)"            "AL-VL1(R)"        
#>  [64] "AL-DM6(R)"         "CX"                "a1(R)"            
#>  [67] "AL-DA3(L)"         "AL-VM7d(L)"        "VES(L)"           
#>  [70] "AL-DL2v(R)"        "AL-VM4(R)"         "RUB(R)"           
#>  [73] "AL-DM5(L)"         "SMP(R)"            "FLA(R)"           
#>  [76] "AL-D(R)"           "AL-VA6(L)"         "NO(R)"            
#>  [79] "FBl6"              "RUB(L)"            "AL-VL2a(R)"       
#>  [82] "g3(R)"             "PED(R)"            "NO1(R)"           
#>  [85] "SPS(L)"            "AL-VA1v(R)"        "PB(L6)"           
#>  [88] "NO(L)"             "AL-DC2(R)"         "AL-DL3(R)"        
#>  [91] "PB(L7)"            "LO(R)"             "NO2(L)"           
#>  [94] "a2(R)"             "AL-DL5(R)"         "AL-DP1l(R)"       
#>  [97] "a'3(R)"            "SMP(L)"            "PENP"             
#> [100] "GA(R)"             "ROB(R)"            "LOP(R)"           
#> [103] "mALT(R)"           "LX(R)"             "SNP(R)"           
#> [106] "AL-DP1m(R)"        "FBl7"              "bL(R)"            
#> [109] "VLNP(R)"           "b'2(R)"            "a'2(R)"           
#> [112] "AL-DM4(L)"         "SNP(L)"            "WED(R)"           
#> [115] "PB(R8)"            "PLP(R)"            "FBl5"             
#> [118] "AL-DC3(R)"         "AL-VM1(R)"         "EPA(R)"           
#> [121] "AL-VM5v(R)"        "AL-DM2(R)"         "AL-VA5(R)"        
#> [124] "AL-VL2p(R)"        "AL-VM2(R)"         "SPS(R)"           
#> [127] "NO3(R)"            "b2(R)"             "AL-DA2(L)"        
#> [130] "CRE(-ROB,-RUB)(R)" "EBr2r4"            "AOTU(R)"          
#> [133] "AL(L)"             "g5(R)"             "FBl3"             
#> [136] "AL-VP5(R)"         "LH(R)"             "AL-VM7d(R)"       
#> [139] "ATL(L)"            "PB(R6)"            "AL-DL4(L)"        
#> [142] "CRE(-RUB)(L)"      "CAN(R)"            "AL-D(L)"          
#> [145] "NO2(R)"            "AL-VC4(R)"         "AL-DL4(R)"        
#> [148] "AL-DM3(R)"         "AL-VP3(R)"         "EBr1"             
#> [151] "dACA(R)"           "PB"                "GNG"              
#> [154] "AL-DA4m(R)"        "aL(R)"             "NO"               
#> [157] "FBl9"              "PB(R2)"            "AL-VC1(R)"        
#> [160] "AL-VC2(R)"         "a'L(R)"            "AL-DM6(L)"        
#> [163] "PB(R9)"            "AL-DA2(R)"         "FB"               
#> [166] "INP"               "OL(R)"             "b'1(R)"           
#> [169] "a3(R)"             "SCL(R)"            "AL-DM1(L)"        
#> [172] "AB(R)"             "GOR(R)"            "AL-VP1d(R)"       
#> [175] "CA(L)"             "VMNP"              "SIP(R)"           
#> [178] "MB(+ACA)(R)"       "AL-DA3(R)"         "PB(R4)"           
#> [181] "PB(R5)"            "AL-DM3(L)"         "CRE(L)"           
#> [184] "LAL(R)"            "AL-VP1m(R)"        "LX(L)"            
#> [187] "AL-VC5(R)"         "AL-VM3(R)"         "SAD"              
#> [190] "PB(R7)"            "EBr6"              "EBr3pw"           
#> [193] "PB(L5)"            "GC"                "AL-DC4(L)"        
#> [196] "PB(R3)"            "EBr3am"            "ICL(L)"           
#> [199] "AL-VP1l(R)"        "EPA(L)"            "AL-VA1d(R)"       
#> [202] "g2(R)"             "EBr3d"             "ATL(R)"           
#> [205] "aL(L)"             "a'L(L)"            "AL-DL1(R)"        
#> [208] "IB"                "BU(L)"             "SAD(-AMMC)"       
#> [211] "PVLP(R)"           "AME(R)"            "AL-VA3(R)"        
#> [214] "AL-V(R)"           "MB(L)"             "AL-DA4m(L)"       
#> [217] "VES(R)"            "FBl1"              "AL-DM2(L)"        
#> [220] "PB(R1)"            "mALT(L)"           "AL-VA6(R)"        
#> [223] "lACA(R)"           "FBl4"              "MB(R)"            
#> [226] "AL-DM4(R)"         "AL-DA1(R)"         "AL-VA2(R)"        
#> [229] "PB(L2)"            "LAL(-GA)(R)"      
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                                     
#> [1,] "GET"     "/api/v:ver/npexplorer/distribution"     
#> [2,] "POST"    "/api/v:ver/roimeshes/mesh/:dataset/:roi"
#> [3,] "OPTIONS" "/api"                                   
#> [4,] "POST"    "/api/raw/keyvalue/key/:instance/:key"   
#> [5,] "POST"    "/api/skeletons/skeleton/:dataset/:id"   
#> [6,] "GET"     "/api/v:ver/roimeshes/mesh/:dataset/:roi"
#>      name                                                                    
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    
#> [3,] "github.com/labstack/echo.glob..func1"                                  
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"
#> [5,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }