Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "CRE(-RUB)(L)"      "FBl4"              "FBl8"             
#>   [4] "ATL(R)"            "AL-DM3(R)"         "g1(R)"            
#>   [7] "FLA(R)"            "AL-DM3(L)"         "AL-VA6(R)"        
#>  [10] "AL-VC2(R)"         "g3(R)"             "FBl2"             
#>  [13] "IPS(R)"            "FBl5"              "AL-DL2d(R)"       
#>  [16] "AL-VA5(R)"         "ROB(R)"            "NO3(R)"           
#>  [19] "AL-VP2(R)"         "EBr3am"            "FB-column3"       
#>  [22] "PB(L4)"            "AL-VP5(R)"         "bL(L)"            
#>  [25] "ICL(R)"            "CRE(R)"            "AL-D(R)"          
#>  [28] "GOR(R)"            "IB112"             "AL-DM1(R)"        
#>  [31] "PB(L8)"            "NO1(R)"            "NO(R)"            
#>  [34] "PENP"              "PB(R8)"            "AL-DL4(L)"        
#>  [37] "a1(R)"             "NO"                "FBl9"             
#>  [40] "AL-VA1d(R)"        "AL-DC1(L)"         "AL-VA7l(R)"       
#>  [43] "AL-VM3(R)"         "MB(L)"             "EBr5"             
#>  [46] "dACA(R)"           "PB(R9)"            "SLP(R)"           
#>  [49] "CRE(-ROB,-RUB)(R)" "VES(L)"            "AL-VA4(R)"        
#>  [52] "AL-DM6(L)"         "FBl3"              "CAN(R)"           
#>  [55] "SMP(R)"            "gL(R)"             "SNP(L)"           
#>  [58] "AL-DA2(R)"         "GA(R)"             "PB(R3)"           
#>  [61] "SCL(L)"            "PLP(R)"            "LX(R)"            
#>  [64] "AL-VM7v(L)"        "g4(R)"             "CX"               
#>  [67] "CA(R)"             "b'L(R)"            "g2(R)"            
#>  [70] "a'2(R)"            "AL-DA4m(L)"        "AL-DM2(L)"        
#>  [73] "AL-DL3(R)"         "AL-VA2(R)"         "AB(L)"            
#>  [76] "EBr6"              "ME(R)"             "EBr3d"            
#>  [79] "AL-DC3(R)"         "AL-VM2(R)"         "LAL(-GA)(R)"      
#>  [82] "EBr3pw"            "AL-DC2(R)"         "AOT(R)"           
#>  [85] "SAD"               "AL-DL1(R)"         "AL-DL4(R)"        
#>  [88] "LX(L)"             "AL-DC1(R)"         "AL-DM4(R)"        
#>  [91] "AL-VL2p(R)"        "LOP(R)"            "SPS(R)"           
#>  [94] "PB"                "g5(R)"             "AL-VA7m(R)"       
#>  [97] "PB(L2)"            "AL-VA3(R)"         "bL(R)"            
#> [100] "AL-VP1m(R)"        "POC"               "RUB(R)"           
#> [103] "ICL(L)"            "FBl6"              "MB(R)"            
#> [106] "AL-DM1(L)"         "AL-VM5v(R)"        "AL-VM7d(L)"       
#> [109] "AL-DA3(R)"         "AL-VC1(R)"         "AL-VM1(R)"        
#> [112] "AL-DA4l(R)"        "PB(L1)"            "FBl7"             
#> [115] "AVLP(R)"           "SCL(R)"            "AL-VC5(R)"        
#> [118] "aL(L)"             "AL-VL2a(R)"        "gL(L)"            
#> [121] "ATL(L)"            "LAL(R)"            "AL-DM2(R)"        
#> [124] "AL-DP1m(R)"        "PVLP(R)"           "PB(R4)"           
#> [127] "AL-DL2v(R)"        "AL-VA6(L)"         "PB(L6)"           
#> [130] "INP"               "FB"                "AL-DA3(L)"        
#> [133] "AL-DL5(L)"         "vACA(R)"           "PRW"              
#> [136] "a3(R)"             "AL-V(R)"           "CA(L)"            
#> [139] "b'2(R)"            "EBr1"              "LH(R)"            
#> [142] "AB(R)"             "PB(L5)"            "EPA(R)"           
#> [145] "AL-DC4(L)"         "VES(R)"            "PED(R)"           
#> [148] "LO(R)"             "AL-DA2(L)"         "AL-DL5(R)"        
#> [151] "b'1(R)"            "AL-DA1(R)"         "AL-VA1v(R)"       
#> [154] "AL-VM7v(R)"        "VLNP(R)"           "GF(R)"            
#> [157] "AL-DM4(L)"         "AL-DM5(L)"         "lACA(R)"          
#> [160] "AL(R)"             "AL-DM6(R)"         "b1(R)"            
#> [163] "AL-DP1l(R)"        "PB(L7)"            "BU(L)"            
#> [166] "BU(R)"             "AL-DC2(L)"         "AL-VC3m(R)"       
#> [169] "PB(R5)"            "LAL(L)"            "AOTU(R)"          
#> [172] "mALT(R)"           "AMMC"              "b2(R)"            
#> [175] "AL(L)"             "AL-D(L)"           "RUB(L)"           
#> [178] "AL-DM5(R)"         "PB(R6)"            "PB(R7)"           
#> [181] "GOR(L)"            "AL-VP4(R)"         "mALT(L)"          
#> [184] "NO2(L)"            "SMP(L)"            "PB(L9)"           
#> [187] "MB(+ACA)(R)"       "AME(R)"            "NO2(R)"           
#> [190] "a'3(R)"            "SIP(L)"            "GC"               
#> [193] "a'L(R)"            "AL-VP1d(R)"        "IB"               
#> [196] "EBr2r4"            "AL-VC3l(R)"        "AL-VM4(R)"        
#> [199] "PB(L3)"            "VMNP"              "AL-VM7d(R)"       
#> [202] "CRE(L)"            "NO(L)"             "AL-DC4(R)"        
#> [205] "AL-VP1l(R)"        "GNG"               "b'L(L)"           
#> [208] "a'L(L)"            "WED(R)"            "NO1(L)"           
#> [211] "EB"                "aL(R)"             "FBl1"             
#> [214] "a2(R)"             "hemibrain"         "AL-DP1m(L)"       
#> [217] "SAD(-AMMC)"        "PB(R2)"            "AL-VC4(R)"        
#> [220] "SNP(R)"            "EPA(L)"            "OL(R)"            
#> [223] "AL-VP3(R)"         "SIP(R)"            "AL-DA4m(R)"       
#> [226] "AL-VL1(R)"         "NO3(L)"            "AL-VM5d(R)"       
#> [229] "SPS(L)"            "PB(R1)"            "a'1(R)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "aL(L)"             "SNP(L)"            "CRE(R)"           
#>   [4] "b'L(R)"            "NO1(R)"            "SMP(R)"           
#>   [7] "SCL(R)"            "EBr6"              "FBl8"             
#>  [10] "NO"                "CRE(L)"            "EBr1"             
#>  [13] "VMNP"              "AME(R)"            "a'L(L)"           
#>  [16] "CRE(-ROB,-RUB)(R)" "GOR(L)"            "aL(R)"            
#>  [19] "NO2(L)"            "SIP(L)"            "EPA(L)"           
#>  [22] "SIP(R)"            "FBl2"              "PB(R2)"           
#>  [25] "PB(R4)"            "b'L(L)"            "SCL(L)"           
#>  [28] "NO3(R)"            "PB"                "bL(L)"            
#>  [31] "b1(R)"             "GOR(R)"            "lACA(R)"          
#>  [34] "LX(R)"             "ICL(L)"            "b'2(R)"           
#>  [37] "FB"                "ATL(L)"            "RUB(L)"           
#>  [40] "a'3(R)"            "PRW"               "PENP"             
#>  [43] "GF(R)"             "BU(L)"             "a'L(R)"           
#>  [46] "LH(R)"             "AVLP(R)"           "RUB(R)"           
#>  [49] "LAL(L)"            "GC"                "FBl9"             
#>  [52] "mALT(R)"           "EBr5"              "b'1(R)"           
#>  [55] "MB(R)"             "PB(R5)"            "g5(R)"            
#>  [58] "ICL(R)"            "LAL(R)"            "PLP(R)"           
#>  [61] "NO(R)"             "EBr2r4"            "gL(L)"            
#>  [64] "SAD(-AMMC)"        "FBl3"              "CAN(R)"           
#>  [67] "AB(L)"             "mALT(L)"           "PB(R6)"           
#>  [70] "ME(R)"             "hemibrain"         "PB(L5)"           
#>  [73] "PB(R8)"            "AL(R)"             "PB(L7)"           
#>  [76] "FBl1"              "VLNP(R)"           "AOTU(R)"          
#>  [79] "BU(R)"             "SMP(L)"            "PB(L6)"           
#>  [82] "CA(L)"             "SPS(L)"            "OL(R)"            
#>  [85] "EBr3am"            "LOP(R)"            "SPS(R)"           
#>  [88] "GNG"               "PB(L8)"            "EBr3pw"           
#>  [91] "PB(R3)"            "AB(R)"             "PED(R)"           
#>  [94] "b2(R)"             "CA(R)"             "INP"              
#>  [97] "NO(L)"             "CX"                "g1(R)"            
#> [100] "FBl5"              "EPA(R)"            "IPS(R)"           
#> [103] "NO2(R)"            "AOT(R)"            "WED(R)"           
#> [106] "EB"                "dACA(R)"           "SNP(R)"           
#> [109] "AL(L)"             "LO(R)"             "VES(R)"           
#> [112] "FBl4"              "SAD"               "MB(L)"            
#> [115] "SLP(R)"            "FBl7"              "LX(L)"            
#> [118] "PB(R9)"            "MB(+ACA)(R)"       "PB(L4)"           
#> [121] "AMMC"              "LAL(-GA)(R)"       "NO1(L)"           
#> [124] "PB(R7)"            "ROB(R)"            "NO3(L)"           
#> [127] "FBl6"              "VES(L)"            "IB"               
#> [130] "PB(L3)"            "GA(R)"             "vACA(R)"          
#> [133] "a1(R)"             "gL(R)"             "ATL(R)"           
#> [136] "PB(L2)"            "a2(R)"             "EBr3d"            
#> [139] "PB(R1)"            "a'1(R)"            "bL(R)"            
#> [142] "a3(R)"             "g2(R)"             "a'2(R)"           
#> [145] "FLA(R)"            "g4(R)"             "PB(L9)"           
#> [148] "PVLP(R)"           "POC"               "g3(R)"            
#> [151] "PB(L1)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "MB(L)"             "AL-DM5(L)"         "AL-VM7v(L)"       
#>   [4] "NO"                "PB(L9)"            "VLNP(R)"          
#>   [7] "CX"                "VMNP"              "g4(R)"            
#>  [10] "CA(R)"             "AL-DA2(R)"         "AL-VC4(R)"        
#>  [13] "GF(R)"             "SIP(R)"            "FBl1"             
#>  [16] "OL(R)"             "PB(L2)"            "lACA(R)"          
#>  [19] "b'L(R)"            "NO1(R)"            "b1(R)"            
#>  [22] "AL-VM2(R)"         "MB(+ACA)(R)"       "AL-VA4(R)"        
#>  [25] "AL-VL2a(R)"        "g1(R)"             "AL-DC2(L)"        
#>  [28] "IPS(R)"            "AL(R)"             "AL-VA3(R)"        
#>  [31] "AB(L)"             "GA(R)"             "PB(L4)"           
#>  [34] "PB(L5)"            "SMP(L)"            "LAL(-GA)(R)"      
#>  [37] "bL(R)"             "AL-VP5(R)"         "PB(L3)"           
#>  [40] "PLP(R)"            "AL-VP2(R)"         "EBr5"             
#>  [43] "SPS(L)"            "AL-DP1m(L)"        "a'1(R)"           
#>  [46] "AVLP(R)"           "AL-VA5(R)"         "AL-VM3(R)"        
#>  [49] "aL(R)"             "LH(R)"             "AL-DC2(R)"        
#>  [52] "AL-DL2v(R)"        "AL-DA3(L)"         "CAN(R)"           
#>  [55] "AL-VP1d(R)"        "mALT(L)"           "EBr3am"           
#>  [58] "GOR(L)"            "AL-VA6(L)"         "AMMC"             
#>  [61] "LX(L)"             "CRE(-ROB,-RUB)(R)" "mALT(R)"          
#>  [64] "LX(R)"             "PVLP(R)"           "b'1(R)"           
#>  [67] "a'3(R)"            "PENP"              "AL-DC1(R)"        
#>  [70] "AL-VM7v(R)"        "g2(R)"             "AL-DM4(R)"        
#>  [73] "AL-DM6(R)"         "FBl3"              "AL-D(R)"          
#>  [76] "AL-VA1v(R)"        "EBr1"              "AL-DL5(L)"        
#>  [79] "NO2(R)"            "a1(R)"             "g3(R)"            
#>  [82] "SAD"               "SCL(R)"            "CRE(L)"           
#>  [85] "SNP(L)"            "NO(L)"             "WED(R)"           
#>  [88] "PB"                "FBl5"              "AL-DL2d(R)"       
#>  [91] "PB(L1)"            "AL-VM7d(L)"        "g5(R)"            
#>  [94] "b'L(L)"            "AL-VC3m(R)"        "IB"               
#>  [97] "GNG"               "AL-VC1(R)"         "SCL(L)"           
#> [100] "FLA(R)"            "AL-DM2(R)"         "AL-VA1d(R)"       
#> [103] "AL-VP4(R)"         "AL-DM3(R)"         "b2(R)"            
#> [106] "AL-DL5(R)"         "AL-VM7d(R)"        "PB(R5)"           
#> [109] "b'2(R)"            "gL(R)"             "FBl2"             
#> [112] "GOR(R)"            "AL-DP1m(R)"        "SMP(R)"           
#> [115] "PED(R)"            "BU(L)"             "FBl8"             
#> [118] "AL-DA1(R)"         "AL-VC2(R)"         "AL-VM1(R)"        
#> [121] "dACA(R)"           "LO(R)"             "EBr6"             
#> [124] "SNP(R)"            "ROB(R)"            "NO(R)"            
#> [127] "FBl9"              "a'L(R)"            "bL(L)"            
#> [130] "AL-DM1(L)"         "FB-column3"        "ATL(R)"           
#> [133] "MB(R)"             "AL-DA4m(R)"        "AL-DM5(R)"        
#> [136] "AL-VA6(R)"         "vACA(R)"           "EB"               
#> [139] "PB(L6)"            "EBr3pw"            "AME(R)"           
#> [142] "AL-DC3(R)"         "AL-DP1l(R)"        "PB(L8)"           
#> [145] "FBl7"              "AL-DL3(R)"         "AL-VA7m(R)"       
#> [148] "AL-VM5d(R)"        "PB(R6)"            "hemibrain"        
#> [151] "AL-DA4m(L)"        "CRE(-RUB)(L)"      "AL-DC4(R)"        
#> [154] "gL(L)"             "AL-DL4(L)"         "AL-VL1(R)"        
#> [157] "AL-V(R)"           "a2(R)"             "NO3(R)"           
#> [160] "EBr3d"             "FBl4"              "AL-VC5(R)"        
#> [163] "POC"               "AL-DM1(R)"         "SPS(R)"           
#> [166] "PB(R3)"            "AL-VM5v(R)"        "LAL(L)"           
#> [169] "EBr2r4"            "PB(R9)"            "a'L(L)"           
#> [172] "SLP(R)"            "NO2(L)"            "AL-VP1l(R)"       
#> [175] "ATL(L)"            "PB(R8)"            "AL-DM6(L)"        
#> [178] "PB(R7)"            "AL-DM2(L)"         "RUB(L)"           
#> [181] "GC"                "FBl6"              "AL-DL4(R)"        
#> [184] "AOT(R)"            "AL(L)"             "AL-VA2(R)"        
#> [187] "AL-VA7l(R)"        "BU(R)"             "AL-DA2(L)"        
#> [190] "AL-D(L)"           "SAD(-AMMC)"        "AL-VP3(R)"        
#> [193] "CRE(R)"            "FB"                "VES(L)"           
#> [196] "PRW"               "ME(R)"             "NO3(L)"           
#> [199] "ICL(R)"            "AL-DA3(R)"         "PB(L7)"           
#> [202] "LOP(R)"            "AOTU(R)"           "RUB(R)"           
#> [205] "PB(R2)"            "AL-VL2p(R)"        "AL-VP1m(R)"       
#> [208] "CA(L)"             "a'2(R)"            "AL-DC4(L)"        
#> [211] "PB(R4)"            "NO1(L)"            "PB(R1)"           
#> [214] "ICL(L)"            "LAL(R)"            "EPA(R)"           
#> [217] "AL-DM4(L)"         "VES(R)"            "INP"              
#> [220] "AL-DL1(R)"         "AL-VM4(R)"         "a3(R)"            
#> [223] "AL-DA4l(R)"        "SIP(L)"            "AL-DC1(L)"        
#> [226] "AL-DM3(L)"         "AB(R)"             "aL(L)"            
#> [229] "EPA(L)"            "AL-VC3l(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "FBl2"              "ATL(L)"            "g5(R)"            
#>   [4] "LH(R)"             "AL-DA3(L)"         "AL-DM4(L)"        
#>   [7] "EPA(R)"            "PB(L5)"            "PB(L3)"           
#>  [10] "CRE(-RUB)(L)"      "AL-VA7l(R)"        "AL-DM4(R)"        
#>  [13] "AB(L)"             "LO(R)"             "AL-DA3(R)"        
#>  [16] "PB(R3)"            "MB(R)"             "AL-DA4m(R)"       
#>  [19] "AL-VC3l(R)"        "AL-VP5(R)"         "RUB(R)"           
#>  [22] "SIP(L)"            "SIP(R)"            "AL-DM5(R)"        
#>  [25] "FB-column3"        "CAN(R)"            "AL-VC2(R)"        
#>  [28] "AL-VC5(R)"         "AL-VM5d(R)"        "POC"              
#>  [31] "LAL(L)"            "NO1(L)"            "PENP"             
#>  [34] "PB(L6)"            "AL-VM7d(L)"        "AL-VM7v(L)"       
#>  [37] "AL-VP3(R)"         "GA(R)"             "FBl1"             
#>  [40] "LAL(-GA)(R)"       "AL-DC3(R)"         "AL-DA2(L)"        
#>  [43] "b'2(R)"            "EBr1"              "AVLP(R)"          
#>  [46] "AL-VP1d(R)"        "AL-DL5(R)"         "AL-VC3m(R)"       
#>  [49] "EBr3am"            "g1(R)"             "AL-DA4l(R)"       
#>  [52] "CRE(-ROB,-RUB)(R)" "ICL(R)"            "NO2(R)"           
#>  [55] "EBr3d"             "AL-VA6(L)"         "bL(R)"            
#>  [58] "AL-DL1(R)"         "PB(R7)"            "FLA(R)"           
#>  [61] "AL-DM1(L)"         "a3(R)"             "RUB(L)"           
#>  [64] "AMMC"              "AL-DC1(L)"         "AL-DC4(L)"        
#>  [67] "VES(L)"            "AL-VM4(R)"         "gL(L)"            
#>  [70] "g4(R)"             "BU(R)"             "NO(R)"            
#>  [73] "AL-DL2d(R)"        "AL-VM2(R)"         "BU(L)"            
#>  [76] "dACA(R)"           "PB(L7)"            "LAL(R)"           
#>  [79] "WED(R)"            "AME(R)"            "AL-VA1v(R)"       
#>  [82] "AL-VL1(R)"         "AL-DA4m(L)"        "PB(R9)"           
#>  [85] "aL(R)"             "CA(R)"             "EPA(L)"           
#>  [88] "AL-DL3(R)"         "AL-VC4(R)"         "vACA(R)"          
#>  [91] "MB(L)"             "PB(L8)"            "NO1(R)"           
#>  [94] "CRE(L)"            "FBl5"              "SMP(L)"           
#>  [97] "lACA(R)"           "PB(R8)"            "PB(R4)"           
#> [100] "AL-VL2a(R)"        "PB(R5)"            "PB(R6)"           
#> [103] "b2(R)"             "SCL(L)"            "AL-DM5(L)"        
#> [106] "PB(R1)"            "AL-DM3(L)"         "NO3(L)"           
#> [109] "FBl6"              "AL-D(R)"           "SAD"              
#> [112] "b'L(R)"            "AOT(R)"            "EBr3pw"           
#> [115] "AL-VA1d(R)"        "ROB(R)"            "AL-DC1(R)"        
#> [118] "PRW"               "PB(L1)"            "AL-VM5v(R)"       
#> [121] "a'3(R)"            "bL(L)"             "OL(R)"            
#> [124] "NO2(L)"            "AL-DA2(R)"         "AL-VA4(R)"        
#> [127] "CRE(R)"            "AL-DM6(L)"         "FBl9"             
#> [130] "AL-VL2p(R)"        "mALT(R)"           "VMNP"             
#> [133] "AL-VP2(R)"         "g3(R)"             "ATL(R)"           
#> [136] "a'1(R)"            "LX(L)"             "AL(R)"            
#> [139] "AL-DC2(R)"         "AL-VC1(R)"         "AL-V(R)"          
#> [142] "g2(R)"             "ME(R)"             "AL-VA3(R)"        
#> [145] "gL(R)"             "PB(L9)"            "IPS(R)"           
#> [148] "AL-DC4(R)"         "AL-VA7m(R)"        "SNP(R)"           
#> [151] "AL-DL4(L)"         "FBl3"              "AL-VA2(R)"        
#> [154] "AL-VM3(R)"         "FB"                "EBr5"             
#> [157] "AL-DC2(L)"         "ICL(L)"            "INP"              
#> [160] "AL-DL2v(R)"        "AL-DM2(R)"         "GF(R)"            
#> [163] "SPS(R)"            "AL-DM6(R)"         "CA(L)"            
#> [166] "AL-VA5(R)"         "a1(R)"             "PVLP(R)"          
#> [169] "AL-DM1(R)"         "a'2(R)"            "AL-D(L)"          
#> [172] "SNP(L)"            "AB(R)"             "PB(L4)"           
#> [175] "SMP(R)"            "EBr6"              "AL-DP1m(L)"       
#> [178] "b'1(R)"            "AL(L)"             "MB(+ACA)(R)"      
#> [181] "NO(L)"             "AL-DP1m(R)"        "LOP(R)"           
#> [184] "a2(R)"             "aL(L)"             "CX"               
#> [187] "LX(R)"             "IB"                "PB"               
#> [190] "SLP(R)"            "hemibrain"         "a'L(R)"           
#> [193] "a'L(L)"            "AL-DP1l(R)"        "NO"               
#> [196] "FBl4"              "AL-VP1l(R)"        "PB(L2)"           
#> [199] "mALT(L)"           "AOTU(R)"           "AL-DL4(R)"        
#> [202] "FBl8"              "PB(R2)"            "AL-VA6(R)"        
#> [205] "AL-VM7d(R)"        "AL-VM7v(R)"        "GC"               
#> [208] "VES(R)"            "AL-DL5(L)"         "GNG"              
#> [211] "b'L(L)"            "NO3(R)"            "SCL(R)"           
#> [214] "AL-VP4(R)"         "GOR(R)"            "AL-DM3(R)"        
#> [217] "SAD(-AMMC)"        "b1(R)"             "AL-VP1m(R)"       
#> [220] "PLP(R)"            "EBr2r4"            "AL-DM2(L)"        
#> [223] "AL-VM1(R)"         "SPS(L)"            "EB"               
#> [226] "FBl7"              "AL-DA1(R)"         "VLNP(R)"          
#> [229] "PED(R)"            "GOR(L)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method  path                                
#> [1,] "POST"  "/api/npexplorer/roisinneuron"      
#> [2,] "POST"  "/api/npexplorer/commonconnectivity"
#> [3,] "POST"  "/api/npexplorer/autapses"          
#> [4,] "TRACE" "/api"                              
#> [5,] "PUT"   "/api/*"                            
#> [6,] "TRACE" "/api/*"                            
#>      name                                                                                              
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"      
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCommonConnectivity-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"          
#> [4,] "github.com/labstack/echo.glob..func1"                                                            
#> [5,] "github.com/labstack/echo.glob..func1"                                                            
#> [6,] "github.com/labstack/echo.glob..func1"                                                            

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }