Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $fib19
#> $fib19$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $fib19$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $fib19$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $fib19$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $fib19$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $fib19$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "WED(R)"            "ME(R)"             "AL-DC3(R)"        
#>   [4] "AL-VA6(R)"         "AL-VC4(R)"         "PB(R7)"           
#>   [7] "RUB(L)"            "AL-VP5(R)"         "AL-VP1m(R)"       
#>  [10] "SMP(R)"            "PED(R)"            "NO1(L)"           
#>  [13] "PB(R8)"            "PVLP(R)"           "ICL(R)"           
#>  [16] "PB(R4)"            "SMP(L)"            "NO3(L)"           
#>  [19] "AL-VP4(R)"         "IB"                "NO2(R)"           
#>  [22] "AL-VA1d(R)"        "AL-VM7v(L)"        "AL-DC1(R)"        
#>  [25] "AL-VL1(R)"         "aL(R)"             "bL(L)"            
#>  [28] "FBl8"              "AL-DM2(R)"         "AL-DP1l(R)"       
#>  [31] "hemibrain"         "g2(R)"             "AL-DA3(L)"        
#>  [34] "AL-VM2(R)"         "FBl1"              "FLA(R)"           
#>  [37] "LX(R)"             "a'2(R)"            "EBr2r4"           
#>  [40] "AL-DM1(L)"         "AL-DM4(L)"         "LO(R)"            
#>  [43] "FBl3"              "AL-DL4(R)"         "AL-DM4(R)"        
#>  [46] "AL-VM5v(R)"        "AL-VM7d(R)"        "LAL(-GA)(R)"      
#>  [49] "NO(R)"             "SAD(-AMMC)"        "CAN(R)"           
#>  [52] "NO(L)"             "PRW"               "AL-DC4(L)"        
#>  [55] "AL-DL1(R)"         "AL-VA7m(R)"        "LAL(R)"           
#>  [58] "b'1(R)"            "AL-DC2(L)"         "FBl2"             
#>  [61] "aL(L)"             "AL-DA4l(R)"        "FB-column3"       
#>  [64] "PB(R2)"            "AL-DM2(L)"         "VLNP(R)"          
#>  [67] "CX"                "gL(R)"             "PB(L5)"           
#>  [70] "PLP(R)"            "EB"                "b1(R)"            
#>  [73] "GNG"               "AL-VC1(R)"         "mALT(R)"          
#>  [76] "SIP(R)"            "FBl9"              "AL-VA1v(R)"       
#>  [79] "AL-VM1(R)"         "vACA(R)"           "PENP"             
#>  [82] "PB(L3)"            "AL-VA6(L)"         "CRE(L)"           
#>  [85] "AL-DM3(R)"         "AL-VA5(R)"         "AL-VL2a(R)"       
#>  [88] "IPS(R)"            "AL-DM6(R)"         "AOT(R)"           
#>  [91] "b'L(L)"            "FBl5"              "b'L(R)"           
#>  [94] "dACA(R)"           "g5(R)"             "a'L(L)"           
#>  [97] "AL-VP1l(R)"        "BU(L)"             "PB(L8)"           
#> [100] "ATL(R)"            "VES(L)"            "AL-VL2p(R)"       
#> [103] "PB(L2)"            "EBr3am"            "g4(R)"            
#> [106] "FBl4"              "AB(R)"             "OL(R)"            
#> [109] "NO3(R)"            "MB(+ACA)(R)"       "AL-VM3(R)"        
#> [112] "SIP(L)"            "PB(L1)"            "NO1(R)"           
#> [115] "PB(L7)"            "AL-VP2(R)"         "EBr3pw"           
#> [118] "AL-DL5(R)"         "LOP(R)"            "AL(R)"            
#> [121] "GC"                "SCL(R)"            "PB(L6)"           
#> [124] "AL-DL2v(R)"        "AL-VP1d(R)"        "AL-VC2(R)"        
#> [127] "SAD"               "AL-DM6(L)"         "AL-VM7v(R)"       
#> [130] "AOTU(R)"           "FB"                "CA(R)"            
#> [133] "EPA(L)"            "lACA(R)"           "PB(R1)"           
#> [136] "AME(R)"            "INP"               "AL-DM5(R)"        
#> [139] "LAL(L)"            "ROB(R)"            "PB(R3)"           
#> [142] "AL-DL2d(R)"        "g1(R)"             "SNP(R)"           
#> [145] "SNP(L)"            "AL-DP1m(R)"        "EBr3d"            
#> [148] "AL-DA4m(L)"        "SLP(R)"            "AL-VM5d(R)"       
#> [151] "MB(L)"             "AL-DA3(R)"         "PB"               
#> [154] "AL-DA2(L)"         "PB(L9)"            "AL-DL3(R)"        
#> [157] "EBr5"              "FBl7"              "MB(R)"            
#> [160] "AL-DC2(R)"         "AL-VM4(R)"         "SCL(L)"           
#> [163] "GA(R)"             "GOR(R)"            "LH(R)"            
#> [166] "a'1(R)"            "FBl6"              "EPA(R)"           
#> [169] "AL-V(R)"           "POC"               "SPS(R)"           
#> [172] "AL-DA1(R)"         "AB(L)"             "b'2(R)"           
#> [175] "GOR(L)"            "AL-DP1m(L)"        "NO"               
#> [178] "mALT(L)"           "NO2(L)"            "CRE(R)"           
#> [181] "AL-DC1(L)"         "AL-VM7d(L)"        "PB(R9)"           
#> [184] "ICL(L)"            "IB112"             "AL-VA4(R)"        
#> [187] "VES(R)"            "AL-D(R)"           "AL-VC3m(R)"       
#> [190] "RUB(R)"            "AMMC"              "g3(R)"            
#> [193] "a'L(R)"            "gL(L)"             "AL-DA4m(R)"       
#> [196] "AL-VA3(R)"         "CA(L)"             "SPS(L)"           
#> [199] "BU(R)"             "AL-VC5(R)"         "PB(R5)"           
#> [202] "a1(R)"             "AL-DL4(L)"         "bL(R)"            
#> [205] "a3(R)"             "AL-VC3l(R)"        "a'3(R)"           
#> [208] "EBr1"              "AL-DM3(L)"         "AL-VP3(R)"        
#> [211] "AL-DM5(L)"         "AL(L)"             "LX(L)"            
#> [214] "CRE(-RUB)(L)"      "PB(L4)"            "ATL(L)"           
#> [217] "b2(R)"             "VMNP"              "AL-VA7l(R)"       
#> [220] "a2(R)"             "AVLP(R)"           "AL-DM1(R)"        
#> [223] "PB(R6)"            "AL-DL5(L)"         "AL-DA2(R)"        
#> [226] "EBr6"              "AL-D(L)"           "CRE(-ROB,-RUB)(R)"
#> [229] "GF(R)"             "AL-DC4(R)"         "AL-VA2(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "SMP(L)"            "NO"                "CX"               
#>   [4] "FB"                "g5(R)"             "FBl7"             
#>   [7] "PB(L8)"            "ME(R)"             "a'3(R)"           
#>  [10] "NO3(R)"            "MB(L)"             "EBr6"             
#>  [13] "lACA(R)"           "FBl9"              "IPS(R)"           
#>  [16] "PB(R5)"            "b'1(R)"            "CRE(-ROB,-RUB)(R)"
#>  [19] "PED(R)"            "g1(R)"             "EB"               
#>  [22] "aL(L)"             "EPA(L)"            "PB(L4)"           
#>  [25] "FLA(R)"            "EBr2r4"            "CAN(R)"           
#>  [28] "SCL(L)"            "vACA(R)"           "LAL(L)"           
#>  [31] "AL(L)"             "EBr3pw"            "CA(R)"            
#>  [34] "a3(R)"             "aL(R)"             "VMNP"             
#>  [37] "FBl2"              "SLP(R)"            "EBr3am"           
#>  [40] "ICL(L)"            "MB(+ACA)(R)"       "ATL(R)"           
#>  [43] "EPA(R)"            "CA(L)"             "NO2(L)"           
#>  [46] "NO(R)"             "FBl1"              "PB(R3)"           
#>  [49] "bL(R)"             "AL(R)"             "AOTU(R)"          
#>  [52] "WED(R)"            "NO1(R)"            "AB(L)"            
#>  [55] "mALT(L)"           "RUB(L)"            "VES(L)"           
#>  [58] "LX(R)"             "EBr3d"             "SNP(L)"           
#>  [61] "NO2(R)"            "PB(L5)"            "b'2(R)"           
#>  [64] "ROB(R)"            "PB(L6)"            "bL(L)"            
#>  [67] "LO(R)"             "FBl8"              "AME(R)"           
#>  [70] "LAL(-GA)(R)"       "SCL(R)"            "BU(R)"            
#>  [73] "gL(R)"             "PENP"              "SPS(L)"           
#>  [76] "PLP(R)"            "b1(R)"             "PB(L3)"           
#>  [79] "PB(R8)"            "PB(L7)"            "a'L(L)"           
#>  [82] "AOT(R)"            "IB"                "PB(L2)"           
#>  [85] "a1(R)"             "SIP(L)"            "PB(R9)"           
#>  [88] "ICL(R)"            "CRE(L)"            "FBl3"             
#>  [91] "LX(L)"             "FBl4"              "INP"              
#>  [94] "GF(R)"             "g3(R)"             "a'1(R)"           
#>  [97] "dACA(R)"           "a'2(R)"            "RUB(R)"           
#> [100] "gL(L)"             "PB(L9)"            "b'L(L)"           
#> [103] "POC"               "ATL(L)"            "hemibrain"        
#> [106] "b2(R)"             "SNP(R)"            "FBl5"             
#> [109] "PB(R2)"            "OL(R)"             "MB(R)"            
#> [112] "PRW"               "PB(R1)"            "SAD(-AMMC)"       
#> [115] "GOR(L)"            "mALT(R)"           "PVLP(R)"          
#> [118] "LOP(R)"            "SAD"               "b'L(R)"           
#> [121] "SIP(R)"            "SMP(R)"            "PB(R7)"           
#> [124] "EBr5"              "a'L(R)"            "GOR(R)"           
#> [127] "LAL(R)"            "SPS(R)"            "VES(R)"           
#> [130] "FBl6"              "LH(R)"             "PB(R4)"           
#> [133] "GNG"               "g2(R)"             "BU(L)"            
#> [136] "g4(R)"             "a2(R)"             "PB(L1)"           
#> [139] "NO(L)"             "NO3(L)"            "PB(R6)"           
#> [142] "GA(R)"             "GC"                "AVLP(R)"          
#> [145] "AB(R)"             "VLNP(R)"           "PB"               
#> [148] "AMMC"              "NO1(L)"            "CRE(R)"           
#> [151] "EBr1"             
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "SMP(L)"            "EBr5"              "SNP(L)"           
#>   [4] "AL-VA1d(R)"        "WED(R)"            "BU(L)"            
#>   [7] "FBl6"              "AL-DL5(R)"         "AL-VL1(R)"        
#>  [10] "VLNP(R)"           "a'2(R)"            "SIP(R)"           
#>  [13] "PB(L6)"            "AL-DM2(R)"         "AL-VM1(R)"        
#>  [16] "RUB(L)"            "PRW"               "AL-DL5(L)"        
#>  [19] "AL-DM1(R)"         "GF(R)"             "g3(R)"            
#>  [22] "ROB(R)"            "CRE(-RUB)(L)"      "b1(R)"            
#>  [25] "AL-VA1v(R)"        "AL-VC5(R)"         "PB(L5)"           
#>  [28] "AL-DC4(L)"         "AL-VA7l(R)"        "AL-VC2(R)"        
#>  [31] "AL-VL2p(R)"        "b2(R)"             "GC"               
#>  [34] "EB"                "AL-DA4m(L)"        "FBl7"             
#>  [37] "AL-VC3m(R)"        "PB(R5)"            "PB(R1)"           
#>  [40] "PB(R2)"            "LAL(R)"            "AL-VA3(R)"        
#>  [43] "EBr3d"             "a1(R)"             "NO2(L)"           
#>  [46] "b'L(L)"            "POC"               "FBl5"             
#>  [49] "IPS(R)"            "AL-VP1d(R)"        "AL-DA3(R)"        
#>  [52] "AL-DC3(R)"         "AL-VC4(R)"         "AL-VP5(R)"        
#>  [55] "AL-DL2v(R)"        "AL-DM3(R)"         "AL-VC3l(R)"       
#>  [58] "PENP"              "AOTU(R)"           "g1(R)"            
#>  [61] "AL-DM1(L)"         "a'1(R)"            "SAD(-AMMC)"       
#>  [64] "GA(R)"             "aL(R)"             "EBr6"             
#>  [67] "AL-DA3(L)"         "CRE(L)"            "PB(R4)"           
#>  [70] "bL(R)"             "AL-V(R)"           "CRE(-ROB,-RUB)(R)"
#>  [73] "SCL(R)"            "ME(R)"             "gL(L)"            
#>  [76] "PB(L9)"            "LO(R)"             "AL-VA2(R)"        
#>  [79] "AL-VM7d(R)"        "AB(L)"             "NO2(R)"           
#>  [82] "LX(L)"             "AL-VP1m(R)"        "EBr1"             
#>  [85] "g5(R)"             "AL-DM4(R)"         "IB"               
#>  [88] "PB(L7)"            "PB(R6)"            "PB(R8)"           
#>  [91] "FBl2"              "PB(R3)"            "AOT(R)"           
#>  [94] "mALT(L)"           "CX"                "PB(R7)"           
#>  [97] "AL-DC2(L)"         "NO(R)"             "bL(L)"            
#> [100] "MB(+ACA)(R)"       "FBl9"              "PLP(R)"           
#> [103] "NO3(L)"            "FBl3"              "MB(R)"            
#> [106] "AL-VM3(R)"         "NO(L)"             "AL-DL3(R)"        
#> [109] "AL-DP1l(R)"        "AL-VM2(R)"         "AL-VP1l(R)"       
#> [112] "LOP(R)"            "a'L(R)"            "AL-DM2(L)"        
#> [115] "AVLP(R)"           "GNG"               "AL-VM5d(R)"       
#> [118] "LAL(-GA)(R)"       "mALT(R)"           "gL(R)"            
#> [121] "AL-DA4l(R)"        "b'1(R)"            "AL-DP1m(L)"       
#> [124] "AL-VA5(R)"         "AL-DM5(R)"         "AL-VC1(R)"        
#> [127] "AL-VM7v(R)"        "a'3(R)"            "LX(R)"            
#> [130] "FB-column3"        "CA(R)"             "LH(R)"            
#> [133] "vACA(R)"           "g2(R)"             "ATL(R)"           
#> [136] "GOR(R)"            "PB(L2)"            "AMMC"             
#> [139] "hemibrain"         "SPS(L)"            "EBr2r4"           
#> [142] "AL-DM4(L)"         "AL-VA6(R)"         "AL-DM6(L)"        
#> [145] "aL(L)"             "AL(R)"             "dACA(R)"          
#> [148] "AL-DM5(L)"         "VES(L)"            "AL-DA4m(R)"       
#> [151] "AL-VM7d(L)"        "AL-DA1(R)"         "AL-DC4(R)"        
#> [154] "AL-VP3(R)"         "AL-DC1(L)"         "AL-DA2(R)"        
#> [157] "FLA(R)"            "SNP(R)"            "BU(R)"            
#> [160] "EBr3am"            "a2(R)"             "AB(R)"            
#> [163] "a3(R)"             "OL(R)"             "AL-DA2(L)"        
#> [166] "AL-DL4(L)"         "ICL(R)"            "ATL(L)"           
#> [169] "b'2(R)"            "SAD"               "VMNP"             
#> [172] "FBl8"              "AL-DC1(R)"         "PB(L3)"           
#> [175] "AL-DL2d(R)"        "SIP(L)"            "AL-DC2(R)"        
#> [178] "PB"                "EPA(L)"            "PB(L4)"           
#> [181] "lACA(R)"           "NO1(L)"            "NO3(R)"           
#> [184] "CRE(R)"            "NO"                "VES(R)"           
#> [187] "PVLP(R)"           "FBl4"              "AME(R)"           
#> [190] "AL-D(R)"           "PB(L8)"            "AL-VL2a(R)"       
#> [193] "AL-VM4(R)"         "SPS(R)"            "AL-D(L)"          
#> [196] "FBl1"              "INP"               "AL-VM5v(R)"       
#> [199] "AL-VP4(R)"         "AL(L)"             "AL-VA7m(R)"       
#> [202] "AL-VP2(R)"         "AL-VM7v(L)"        "MB(L)"            
#> [205] "AL-DP1m(R)"        "CAN(R)"            "AL-DM6(R)"        
#> [208] "FB"                "NO1(R)"            "AL-DM3(L)"        
#> [211] "AL-DL4(R)"         "AL-VA4(R)"         "RUB(R)"           
#> [214] "SCL(L)"            "CA(L)"             "PED(R)"           
#> [217] "PB(L1)"            "g4(R)"             "SMP(R)"           
#> [220] "GOR(L)"            "PB(R9)"            "ICL(L)"           
#> [223] "a'L(L)"            "AL-DL1(R)"         "LAL(L)"           
#> [226] "b'L(R)"            "EPA(R)"            "AL-VA6(L)"        
#> [229] "EBr3pw"            "SLP(R)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-DM1(R)"         "MB(+ACA)(R)"       "AL-DL2v(R)"       
#>   [4] "AL-DA2(L)"         "AL-DL5(L)"         "ATL(R)"           
#>   [7] "AL-VA1d(R)"        "AL-VA6(R)"         "AL-VC4(R)"        
#>  [10] "AL-VL1(R)"         "AL-DA3(L)"         "SIP(R)"           
#>  [13] "FBl3"              "AL-D(R)"           "AL-VA3(R)"        
#>  [16] "AL-VP1d(R)"        "RUB(R)"            "AOTU(R)"          
#>  [19] "FB"                "g5(R)"             "AL-VP2(R)"        
#>  [22] "bL(R)"             "GOR(R)"            "PB(L4)"           
#>  [25] "lACA(R)"           "LH(R)"             "GNG"              
#>  [28] "AOT(R)"            "b'1(R)"            "AL-VA6(L)"        
#>  [31] "bL(L)"             "IB"                "AL-VM2(R)"        
#>  [34] "GA(R)"             "AL-VM7v(L)"        "AL-V(R)"          
#>  [37] "FBl2"              "PB(R4)"            "AMMC"             
#>  [40] "VES(L)"            "AL-VM5d(R)"        "AME(R)"           
#>  [43] "AL-DA2(R)"         "RUB(L)"            "AL-DA4m(R)"       
#>  [46] "EBr2r4"            "AL-DM4(L)"         "LX(R)"            
#>  [49] "AL-DC4(R)"         "AL-VC3l(R)"        "PB(L3)"           
#>  [52] "SIP(L)"            "OL(R)"             "AL-DL3(R)"        
#>  [55] "SMP(R)"            "g3(R)"             "g4(R)"            
#>  [58] "FBl5"              "SCL(L)"            "CRE(-ROB,-RUB)(R)"
#>  [61] "BU(L)"             "AL-DL4(L)"         "PVLP(R)"          
#>  [64] "AL-DL4(R)"         "b'L(R)"            "gL(L)"            
#>  [67] "gL(R)"             "PB(L5)"            "b'2(R)"           
#>  [70] "mALT(R)"           "b2(R)"             "AL-DA4m(L)"       
#>  [73] "FBl1"              "AL-DL1(R)"         "AL-VA4(R)"        
#>  [76] "SPS(R)"            "CRE(R)"            "AL-VP1l(R)"       
#>  [79] "a'L(R)"            "AL-DM2(L)"         "SPS(L)"           
#>  [82] "PB(R9)"            "a3(R)"             "AL-DM4(R)"        
#>  [85] "AL-VA7m(R)"        "SAD"               "EBr1"             
#>  [88] "AL-VA5(R)"         "AL-VM7d(L)"        "AL(R)"            
#>  [91] "g2(R)"             "EB"                "aL(L)"            
#>  [94] "WED(R)"            "NO1(L)"            "AL-DC4(L)"        
#>  [97] "PB(R7)"            "EBr3d"             "AL-DP1m(L)"       
#> [100] "SAD(-AMMC)"        "AL-VP1m(R)"        "CX"               
#> [103] "dACA(R)"           "BU(R)"             "AL-VP5(R)"        
#> [106] "SLP(R)"            "a'3(R)"            "AL-VP4(R)"        
#> [109] "LOP(R)"            "NO(R)"             "AL-DA1(R)"        
#> [112] "PB(L1)"            "AL-DC1(L)"         "FBl7"             
#> [115] "AL-VL2a(R)"        "ME(R)"             "PB(R1)"           
#> [118] "ROB(R)"            "AL-VL2p(R)"        "EBr3am"           
#> [121] "PED(R)"            "SNP(R)"            "PRW"              
#> [124] "ICL(L)"            "EPA(L)"            "AL-DP1m(R)"       
#> [127] "NO"                "b1(R)"             "AL-DC2(L)"        
#> [130] "AL-VA2(R)"         "a'2(R)"            "AL(L)"            
#> [133] "AL-DM5(R)"         "AL-VP3(R)"         "PB"               
#> [136] "PLP(R)"            "PB(L2)"            "NO3(L)"           
#> [139] "FBl8"              "FBl6"              "AL-VA1v(R)"       
#> [142] "AL-VA7l(R)"        "vACA(R)"           "FLA(R)"           
#> [145] "AL-DM1(L)"         "VES(R)"            "PB(R6)"           
#> [148] "a2(R)"             "SNP(L)"            "AL-VC5(R)"        
#> [151] "g1(R)"             "AL-DM5(L)"         "EBr5"             
#> [154] "AL-DM3(L)"         "LX(L)"             "POC"              
#> [157] "PENP"              "GC"                "PB(R3)"           
#> [160] "AB(R)"             "SMP(L)"            "NO3(R)"           
#> [163] "mALT(L)"           "LAL(-GA)(R)"       "AL-VM3(R)"        
#> [166] "MB(R)"             "AL-VC2(R)"         "VMNP"             
#> [169] "FBl9"              "PB(L7)"            "AL-DL5(R)"        
#> [172] "AL-DM6(R)"         "AL-VM5v(R)"        "NO2(L)"           
#> [175] "aL(R)"             "NO1(R)"            "PB(R2)"           
#> [178] "NO2(R)"            "AL-DC3(R)"         "GOR(L)"           
#> [181] "EPA(R)"            "ATL(L)"            "hemibrain"        
#> [184] "AL-VM7d(R)"        "ICL(R)"            "AL-DA4l(R)"       
#> [187] "AL-DP1l(R)"        "AL-VC1(R)"         "AL-VC3m(R)"       
#> [190] "CA(L)"             "EBr6"              "EBr3pw"           
#> [193] "AB(L)"             "INP"               "a'L(L)"           
#> [196] "PB(R5)"            "PB(L9)"            "CAN(R)"           
#> [199] "IPS(R)"            "LAL(R)"            "AL-DL2d(R)"       
#> [202] "AL-VM1(R)"         "MB(L)"             "CA(R)"            
#> [205] "AL-DC2(R)"         "LO(R)"             "AL-VM4(R)"        
#> [208] "AL-DA3(R)"         "SCL(R)"            "PB(L8)"           
#> [211] "AL-D(L)"           "VLNP(R)"           "a1(R)"            
#> [214] "CRE(L)"            "NO(L)"             "FBl4"             
#> [217] "AL-DC1(R)"         "AVLP(R)"           "AL-DM2(R)"        
#> [220] "AL-VM7v(R)"        "LAL(L)"            "GF(R)"            
#> [223] "a'1(R)"            "FB-column3"        "b'L(L)"           
#> [226] "PB(R8)"            "CRE(-RUB)(L)"      "AL-DM3(R)"        
#> [229] "AL-DM6(L)"         "PB(L6)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method   path                                    
#> [1,] "GET"    "/login"                                
#> [2,] "PATCH"  "/api"                                  
#> [3,] "DELETE" "/api/*"                                
#> [4,] "GET"    "/api/v:ver/dbmeta/datasets"            
#> [5,] "POST"   "/api/raw/cypher/transaction/:id/cypher"
#> [6,] "GET"    "/api/v:ver/npexplorer/findneurons"     
#>      name                                                                    
#> [1,] "github.com/janelia-flyem/echo-secure.loginHandler"                     
#> [2,] "github.com/labstack/echo/v4.glob..func1"                               
#> [3,] "github.com/labstack/echo/v4.glob..func1"                               
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    
#> [5,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }