Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "NO3(R)"            "EB"                "AL-VA7m(R)"       
#>   [4] "AOT(R)"            "VMNP"              "AL-DM2(L)"        
#>   [7] "GOR(R)"            "AL-VC4(R)"         "AL-DM4(L)"        
#>  [10] "AL-DP1l(R)"        "CAN(R)"            "AL-VM7v(R)"       
#>  [13] "g4(R)"             "AL-DM1(L)"         "NO2(R)"           
#>  [16] "SLP(R)"            "AL-DC1(R)"         "WED(R)"           
#>  [19] "g1(R)"             "NO3(L)"            "GF(R)"            
#>  [22] "PB(R2)"            "AL(R)"             "AL-VP1d(R)"       
#>  [25] "a2(R)"             "PB(R8)"            "AL-DC1(L)"        
#>  [28] "AL-DL2d(R)"        "PB(R3)"            "AL-DA1(R)"        
#>  [31] "AL-DM3(R)"         "AL-VC5(R)"         "MB(+ACA)(R)"      
#>  [34] "AL-D(L)"           "FBl5"              "a3(R)"            
#>  [37] "NO(L)"             "PB(L3)"            "a'2(R)"           
#>  [40] "AL-DA2(L)"         "AL-DC4(L)"         "AL-DM2(R)"        
#>  [43] "AL-VL1(R)"         "a1(R)"             "AB(L)"            
#>  [46] "SCL(L)"            "GOR(L)"            "EBr1"             
#>  [49] "LH(R)"             "AL-DL4(R)"         "BU(R)"            
#>  [52] "AL-DA4l(R)"        "AL-VA2(R)"         "POC"              
#>  [55] "PB(R6)"            "a'L(R)"            "LO(R)"            
#>  [58] "CX"                "SIP(L)"            "gL(L)"            
#>  [61] "AL-DC4(R)"         "AL-VM5d(R)"        "NO"               
#>  [64] "RUB(L)"            "AL-VA7l(R)"        "mALT(L)"          
#>  [67] "AL-DM4(R)"         "AL-D(R)"           "AL-VA1v(R)"       
#>  [70] "AL-DA4m(L)"        "AB(R)"             "AL-VP4(R)"        
#>  [73] "AMMC"              "b'2(R)"            "AOTU(R)"          
#>  [76] "LX(R)"             "AL-DL5(R)"         "FBl4"             
#>  [79] "EPA(L)"            "PB(R7)"            "mALT(R)"          
#>  [82] "NO(R)"             "SNP(L)"            "AL-VL2p(R)"       
#>  [85] "vACA(R)"           "NO1(R)"            "aL(L)"            
#>  [88] "AL-VP5(R)"         "ICL(L)"            "AL-DC3(R)"        
#>  [91] "AL-DL1(R)"         "AL-DM6(R)"         "AL-VL2a(R)"       
#>  [94] "CA(L)"             "PB(L1)"            "SIP(R)"           
#>  [97] "CRE(-RUB)(L)"      "PB(R4)"            "CRE(R)"           
#> [100] "gL(R)"             "PB(L9)"            "IPS(R)"           
#> [103] "FLA(R)"            "EPA(R)"            "ROB(R)"           
#> [106] "SAD"               "g2(R)"             "AL-DA2(R)"        
#> [109] "AL-VC3m(R)"        "b'L(R)"            "AL-VA3(R)"        
#> [112] "AL-VA4(R)"         "LAL(L)"            "ICL(R)"           
#> [115] "bL(R)"             "AL-DL3(R)"         "AL-DP1m(R)"       
#> [118] "AL-VC2(R)"         "PENP"              "SMP(L)"           
#> [121] "AL-VM7v(L)"        "g5(R)"             "MB(R)"            
#> [124] "SNP(R)"            "FBl6"              "LOP(R)"           
#> [127] "PB(L5)"            "AL-DA3(L)"         "EBr3pw"           
#> [130] "PB(R5)"            "IB"                "EBr2r4"           
#> [133] "a'1(R)"            "SMP(R)"            "AL-VM7d(R)"       
#> [136] "SCL(R)"            "AL-DL5(L)"         "FB-column3"       
#> [139] "AL-VM4(R)"         "CA(R)"             "AL-VP1l(R)"       
#> [142] "RUB(R)"            "PED(R)"            "PB(L6)"           
#> [145] "bL(L)"             "FBl3"              "PB"               
#> [148] "AL-DM6(L)"         "LX(L)"             "GC"               
#> [151] "VLNP(R)"           "EBr3d"             "VES(R)"           
#> [154] "PB(L4)"            "AL-VM5v(R)"        "EBr3am"           
#> [157] "hemibrain"         "SPS(R)"            "AL-DL2v(R)"       
#> [160] "AL-VP2(R)"         "g3(R)"             "AL(L)"            
#> [163] "AL-VA6(L)"         "IB112"             "AL-VM7d(L)"       
#> [166] "PVLP(R)"           "CRE(-ROB,-RUB)(R)" "lACA(R)"          
#> [169] "aL(R)"             "b2(R)"             "AL-DM3(L)"        
#> [172] "PB(R9)"            "AME(R)"            "VES(L)"           
#> [175] "AL-DC2(R)"         "a'L(L)"            "NO1(L)"           
#> [178] "PLP(R)"            "CRE(L)"            "INP"              
#> [181] "NO2(L)"            "PB(L8)"            "SPS(L)"           
#> [184] "PB(R1)"            "b1(R)"             "AL-VM2(R)"        
#> [187] "a'3(R)"            "ATL(L)"            "BU(L)"            
#> [190] "ATL(R)"            "AL-VP1m(R)"        "b'1(R)"           
#> [193] "FBl2"              "FBl7"              "AVLP(R)"          
#> [196] "AL-VC3l(R)"        "FBl8"              "AL-VM1(R)"        
#> [199] "LAL(-GA)(R)"       "FB"                "EBr6"             
#> [202] "ME(R)"             "GA(R)"             "dACA(R)"          
#> [205] "AL-DA4m(R)"        "AL-VA6(R)"         "AL-VC1(R)"        
#> [208] "AL-DM5(L)"         "AL-DP1m(L)"        "SAD(-AMMC)"       
#> [211] "AL-DM5(R)"         "FBl1"              "AL-VM3(R)"        
#> [214] "AL-VA5(R)"         "EBr5"              "b'L(L)"           
#> [217] "AL-DA3(R)"         "AL-DM1(R)"         "AL-DC2(L)"        
#> [220] "OL(R)"             "MB(L)"             "PB(L7)"           
#> [223] "GNG"               "LAL(R)"            "FBl9"             
#> [226] "AL-VP3(R)"         "PB(L2)"            "PRW"              
#> [229] "AL-DL4(L)"         "AL-VA1d(R)"        "AL-V(R)"          
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "AOTU(R)"           "NO(L)"             "POC"              
#>   [4] "PB(L5)"            "NO2(L)"            "EPA(R)"           
#>   [7] "ROB(R)"            "vACA(R)"           "mALT(L)"          
#>  [10] "WED(R)"            "EBr5"              "PB(R3)"           
#>  [13] "AVLP(R)"           "LX(L)"             "LAL(L)"           
#>  [16] "PB"                "gL(L)"             "PVLP(R)"          
#>  [19] "GOR(R)"            "LOP(R)"            "a2(R)"            
#>  [22] "EBr1"              "RUB(L)"            "SMP(R)"           
#>  [25] "IB"                "CRE(L)"            "MB(+ACA)(R)"      
#>  [28] "LAL(R)"            "PB(R5)"            "GF(R)"            
#>  [31] "NO(R)"             "LH(R)"             "CX"               
#>  [34] "lACA(R)"           "SNP(R)"            "FB"               
#>  [37] "bL(R)"             "SLP(R)"            "VLNP(R)"          
#>  [40] "ATL(L)"            "LAL(-GA)(R)"       "NO3(R)"           
#>  [43] "g5(R)"             "g3(R)"             "b2(R)"            
#>  [46] "GC"                "g1(R)"             "aL(R)"            
#>  [49] "dACA(R)"           "FLA(R)"            "VES(R)"           
#>  [52] "GA(R)"             "mALT(R)"           "EBr3am"           
#>  [55] "FBl5"              "NO2(R)"            "NO"               
#>  [58] "PB(L6)"            "PB(L9)"            "FBl6"             
#>  [61] "ATL(R)"            "MB(R)"             "PRW"              
#>  [64] "ME(R)"             "NO3(L)"            "FBl2"             
#>  [67] "CRE(R)"            "g4(R)"             "AMMC"             
#>  [70] "PB(L8)"            "CAN(R)"            "CA(R)"            
#>  [73] "PB(L1)"            "FBl9"              "GNG"              
#>  [76] "PB(R7)"            "AB(R)"             "EPA(L)"           
#>  [79] "b'L(L)"            "a'L(L)"            "OL(R)"            
#>  [82] "PB(R9)"            "FBl4"              "INP"              
#>  [85] "AOT(R)"            "SPS(R)"            "a'2(R)"           
#>  [88] "SNP(L)"            "aL(L)"             "SMP(L)"           
#>  [91] "a'1(R)"            "PB(R1)"            "bL(L)"            
#>  [94] "FBl3"              "a'3(R)"            "SIP(L)"           
#>  [97] "PLP(R)"            "SCL(R)"            "LX(R)"            
#> [100] "EB"                "SIP(R)"            "PB(R2)"           
#> [103] "b1(R)"             "b'2(R)"            "g2(R)"            
#> [106] "AL(R)"             "b'1(R)"            "EBr3d"            
#> [109] "gL(R)"             "ICL(L)"            "ICL(R)"           
#> [112] "IPS(R)"            "PB(L4)"            "BU(L)"            
#> [115] "VMNP"              "VES(L)"            "EBr2r4"           
#> [118] "BU(R)"             "FBl8"              "a1(R)"            
#> [121] "MB(L)"             "NO1(R)"            "EBr3pw"           
#> [124] "AB(L)"             "RUB(R)"            "NO1(L)"           
#> [127] "a'L(R)"            "GOR(L)"            "AL(L)"            
#> [130] "SAD(-AMMC)"        "PB(R4)"            "CRE(-ROB,-RUB)(R)"
#> [133] "hemibrain"         "PED(R)"            "FBl1"             
#> [136] "AME(R)"            "PB(L2)"            "PB(L3)"           
#> [139] "SAD"               "PB(R6)"            "SCL(L)"           
#> [142] "b'L(R)"            "SPS(L)"            "LO(R)"            
#> [145] "FBl7"              "a3(R)"             "CA(L)"            
#> [148] "PB(R8)"            "PB(L7)"            "PENP"             
#> [151] "EBr6"             
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "FBl2"              "FBl5"              "AL-VA1d(R)"       
#>   [4] "b'L(R)"            "GOR(R)"            "b2(R)"            
#>   [7] "CRE(-ROB,-RUB)(R)" "AL-VM7d(L)"        "SIP(R)"           
#>  [10] "AME(R)"            "PB(L4)"            "AL-DA2(L)"        
#>  [13] "AL-D(R)"           "AL-VM4(R)"         "FBl6"             
#>  [16] "g4(R)"             "BU(R)"             "gL(R)"            
#>  [19] "AL-VP2(R)"         "EBr1"              "mALT(R)"          
#>  [22] "PB(L8)"            "AL(R)"             "AL-VM1(R)"        
#>  [25] "LOP(R)"            "PB(L3)"            "SPS(L)"           
#>  [28] "AL-VC3l(R)"        "AL-VL2p(R)"        "PB(L2)"           
#>  [31] "NO1(R)"            "ATL(L)"            "FLA(R)"           
#>  [34] "b'L(L)"            "AL-DC2(R)"         "AL-DM4(R)"        
#>  [37] "ICL(L)"            "aL(R)"             "AL-DM3(L)"        
#>  [40] "AL-VA4(R)"         "PED(R)"            "AL-DM6(R)"        
#>  [43] "a3(R)"             "AL-D(L)"           "AL-DC2(L)"        
#>  [46] "NO"                "AL-DA3(R)"         "SNP(R)"           
#>  [49] "AL-DC1(L)"         "AL-VP3(R)"         "g2(R)"            
#>  [52] "lACA(R)"           "MB(L)"             "EBr2r4"           
#>  [55] "AL-DM1(L)"         "a'1(R)"            "PB(L7)"           
#>  [58] "FBl8"              "g3(R)"             "AB(L)"            
#>  [61] "bL(R)"             "AL-DL4(L)"         "AL-DC1(R)"        
#>  [64] "AL-DM5(R)"         "AL-VA5(R)"         "EBr5"             
#>  [67] "AL-DA2(R)"         "SMP(R)"            "PB(R2)"           
#>  [70] "AL-DC4(R)"         "AL-DL1(R)"         "AL-DM2(L)"        
#>  [73] "AL-VC3m(R)"        "AL-VM2(R)"         "EBr3am"           
#>  [76] "AL-VA6(L)"         "PB(R7)"            "b'2(R)"           
#>  [79] "VES(R)"            "GNG"               "AL-VA7l(R)"       
#>  [82] "AL-VL2a(R)"        "VMNP"              "CRE(L)"           
#>  [85] "LO(R)"             "a'3(R)"            "CA(L)"            
#>  [88] "AL-VM7v(L)"        "LX(R)"             "INP"              
#>  [91] "SCL(R)"            "AL-VP5(R)"         "AVLP(R)"          
#>  [94] "AL-DC3(R)"         "AL-VM7v(R)"        "g5(R)"            
#>  [97] "SNP(L)"            "aL(L)"             "CRE(R)"           
#> [100] "dACA(R)"           "AL-DA4m(L)"        "AL-DM3(R)"        
#> [103] "SAD"               "NO(R)"             "PB(L9)"           
#> [106] "EPA(L)"            "AL-DL3(R)"         "AL-V(R)"          
#> [109] "PB(L1)"            "NO3(R)"            "AL-DL2d(R)"       
#> [112] "vACA(R)"           "GOR(L)"            "FBl7"             
#> [115] "FBl9"              "b'1(R)"            "NO1(L)"           
#> [118] "bL(L)"             "FBl1"              "IPS(R)"           
#> [121] "mALT(L)"           "a'2(R)"            "AL-DL5(R)"        
#> [124] "AL-DM1(R)"         "EBr3pw"            "a'L(R)"           
#> [127] "AL-DA1(R)"         "GA(R)"             "ME(R)"            
#> [130] "FBl4"              "GF(R)"             "PB(R1)"           
#> [133] "AB(R)"             "AL-VA6(R)"         "CX"               
#> [136] "PENP"              "EBr6"              "PB(L6)"           
#> [139] "AL-VA2(R)"         "SMP(L)"            "NO3(L)"           
#> [142] "CA(R)"             "AL-DL2v(R)"        "SCL(L)"           
#> [145] "GC"                "SLP(R)"            "AL-VP4(R)"        
#> [148] "PB(L5)"            "ATL(R)"            "PB"               
#> [151] "EB"                "CRE(-RUB)(L)"      "AL-DM2(R)"        
#> [154] "AL-VA3(R)"         "LAL(-GA)(R)"       "AL-VA1v(R)"       
#> [157] "VLNP(R)"           "AL-DA3(L)"         "BU(L)"            
#> [160] "a'L(L)"            "AL-VL1(R)"         "AL-VM5v(R)"       
#> [163] "SPS(R)"            "PRW"               "FB"               
#> [166] "b1(R)"             "LAL(R)"            "AL-DA4m(R)"       
#> [169] "AL-DL4(R)"         "PB(R6)"            "AL-DP1m(R)"       
#> [172] "a1(R)"             "MB(R)"             "MB(+ACA)(R)"      
#> [175] "LH(R)"             "AL-VP1l(R)"        "NO2(L)"           
#> [178] "AL-DM4(L)"         "PB(R9)"            "AL-VC4(R)"        
#> [181] "AL-VP1d(R)"        "POC"               "AL-DP1m(L)"       
#> [184] "AL-VM3(R)"         "AL-VM5d(R)"        "AL-VC1(R)"        
#> [187] "AL-VC2(R)"         "AL-DP1l(R)"        "WED(R)"           
#> [190] "AL(L)"             "ICL(R)"            "RUB(L)"           
#> [193] "OL(R)"             "IB"                "a2(R)"            
#> [196] "AL-DL5(L)"         "AL-DC4(L)"         "g1(R)"            
#> [199] "PB(R8)"            "ROB(R)"            "hemibrain"        
#> [202] "AL-VA7m(R)"        "AL-VC5(R)"         "gL(L)"            
#> [205] "PLP(R)"            "AOT(R)"            "AMMC"             
#> [208] "LAL(L)"            "NO2(R)"            "PB(R4)"           
#> [211] "AL-VP1m(R)"        "AL-DM5(L)"         "RUB(R)"           
#> [214] "EPA(R)"            "PVLP(R)"           "FBl3"             
#> [217] "LX(L)"             "PB(R5)"            "AOTU(R)"          
#> [220] "AL-DM6(L)"         "PB(R3)"            "CAN(R)"           
#> [223] "NO(L)"             "EBr3d"             "AL-VM7d(R)"       
#> [226] "VES(L)"            "SAD(-AMMC)"        "FB-column3"       
#> [229] "AL-DA4l(R)"        "SIP(L)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "SIP(L)"            "NO1(L)"            "AL-DC1(R)"        
#>   [4] "FBl3"              "LH(R)"             "FBl6"             
#>   [7] "AL-DL5(R)"         "AL-DP1l(R)"        "AL-VP1l(R)"       
#>  [10] "AB(L)"             "EBr1"              "AL-DA4m(L)"       
#>  [13] "NO3(R)"            "AL-VA7l(R)"        "PVLP(R)"          
#>  [16] "CAN(R)"            "MB(+ACA)(R)"       "AL-VA2(R)"        
#>  [19] "AL-VP4(R)"         "b'2(R)"            "NO3(L)"           
#>  [22] "bL(R)"             "CRE(R)"            "PB(L6)"           
#>  [25] "AL-DC1(L)"         "AL-DL5(L)"         "AL-DM2(L)"        
#>  [28] "IPS(R)"            "SMP(L)"            "BU(L)"            
#>  [31] "AL-DL2d(R)"        "AL-DM2(R)"         "AL-VM1(R)"        
#>  [34] "a'1(R)"            "AVLP(R)"           "INP"              
#>  [37] "AL-VP5(R)"         "SNP(L)"            "PRW"              
#>  [40] "PB(R9)"            "NO(R)"             "AL-VA5(R)"        
#>  [43] "LAL(L)"            "g2(R)"             "LX(R)"            
#>  [46] "AL-DA2(R)"         "GA(R)"             "g1(R)"            
#>  [49] "AL-VM7v(L)"        "AL-VC1(R)"         "ATL(R)"           
#>  [52] "AL-DL4(R)"         "AL-VA4(R)"         "CA(L)"            
#>  [55] "NO1(R)"            "AL-D(L)"           "FBl8"             
#>  [58] "PB(R2)"            "EBr6"              "SPS(L)"           
#>  [61] "LOP(R)"            "SNP(R)"            "AL-VL1(R)"        
#>  [64] "AL-DA3(R)"         "AL-DP1m(R)"        "AL-VC3m(R)"       
#>  [67] "b1(R)"             "PB(L4)"            "AL-DM1(L)"        
#>  [70] "FBl9"              "OL(R)"             "MB(L)"            
#>  [73] "AL(L)"             "AL-VC2(R)"         "VES(R)"           
#>  [76] "AL-D(R)"           "CA(R)"             "AL-VM4(R)"        
#>  [79] "GC"                "PB"                "ICL(L)"           
#>  [82] "b'L(L)"            "AL-DA4l(R)"        "AL-DC4(R)"        
#>  [85] "IB"                "SAD"               "b2(R)"            
#>  [88] "PB(L2)"            "aL(R)"             "AL-DM3(R)"        
#>  [91] "MB(R)"             "a3(R)"             "AL(R)"            
#>  [94] "AL-DC2(R)"         "ME(R)"             "FBl2"             
#>  [97] "ROB(R)"            "PB(R7)"            "AL-DA3(L)"        
#> [100] "FBl4"              "RUB(L)"            "PB(R4)"           
#> [103] "AL-VC4(R)"         "AL-VL2a(R)"        "SPS(R)"           
#> [106] "gL(R)"             "SLP(R)"            "AL-VC3l(R)"       
#> [109] "VLNP(R)"           "EBr3am"            "PENP"             
#> [112] "AME(R)"            "AB(R)"             "AL-DM6(R)"        
#> [115] "CRE(-ROB,-RUB)(R)" "PB(R6)"            "PB(R3)"           
#> [118] "AL-DA2(L)"         "AL-VM2(R)"         "PLP(R)"           
#> [121] "EBr2r4"            "PB(L1)"            "AL-VL2p(R)"       
#> [124] "AL-VP1m(R)"        "PB(R5)"            "b'1(R)"           
#> [127] "hemibrain"         "PB(L8)"            "AL-DA1(R)"        
#> [130] "FBl7"              "AL-DA4m(R)"        "AL-VM3(R)"        
#> [133] "AL-VM7v(R)"        "AL-VP2(R)"         "EPA(R)"           
#> [136] "bL(L)"             "SCL(L)"            "GF(R)"            
#> [139] "AL-DC3(R)"         "AL-DM5(R)"         "gL(L)"            
#> [142] "AL-DL1(R)"         "aL(L)"             "b'L(R)"           
#> [145] "FBl1"              "AL-DM4(L)"         "AL-VA6(L)"        
#> [148] "PB(L9)"            "g4(R)"             "AL-DL4(L)"        
#> [151] "PB(L5)"            "AMMC"              "mALT(R)"          
#> [154] "PB(R8)"            "PB(R1)"            "AL-DL2v(R)"       
#> [157] "mALT(L)"           "EBr5"              "dACA(R)"          
#> [160] "FLA(R)"            "SAD(-AMMC)"        "AL-VA6(R)"        
#> [163] "CX"                "AL-DP1m(L)"        "FB-column3"       
#> [166] "RUB(R)"            "GOR(L)"            "EBr3d"            
#> [169] "GOR(R)"            "LAL(R)"            "g3(R)"            
#> [172] "EPA(L)"            "LAL(-GA)(R)"       "AL-VM7d(L)"       
#> [175] "NO"                "AOTU(R)"           "PED(R)"           
#> [178] "FBl5"              "AOT(R)"            "lACA(R)"          
#> [181] "SIP(R)"            "AL-DC2(L)"         "a'L(L)"           
#> [184] "AL-DM1(R)"         "AL-DM5(L)"         "EBr3pw"           
#> [187] "CRE(-RUB)(L)"      "AL-VM5v(R)"        "AL-V(R)"          
#> [190] "VMNP"              "AL-DM6(L)"         "AL-VM7d(R)"       
#> [193] "AL-VP3(R)"         "ICL(R)"            "AL-DM4(R)"        
#> [196] "PB(L7)"            "AL-VA1v(R)"        "AL-DC4(L)"        
#> [199] "GNG"               "AL-VA7m(R)"        "ATL(L)"           
#> [202] "a2(R)"             "WED(R)"            "AL-VC5(R)"        
#> [205] "NO2(L)"            "FB"                "AL-VA1d(R)"       
#> [208] "a'2(R)"            "VES(L)"            "SCL(R)"           
#> [211] "AL-VM5d(R)"        "BU(R)"             "vACA(R)"          
#> [214] "AL-VP1d(R)"        "NO2(R)"            "NO(L)"            
#> [217] "CRE(L)"            "LO(R)"             "AL-DL3(R)"        
#> [220] "POC"               "PB(L3)"            "EB"               
#> [223] "a'3(R)"            "g5(R)"             "AL-VA3(R)"        
#> [226] "a'L(R)"            "SMP(R)"            "a1(R)"            
#> [229] "AL-DM3(L)"         "LX(L)"            
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method     path                                    
#> [1,] "REPORT"   "/api"                                  
#> [2,] "OPTIONS"  "/api"                                  
#> [3,] "PROPFIND" "/api"                                  
#> [4,] "POST"     "/api/raw/cypher/transaction/:id/cypher"
#> [5,] "POST"     "/api/v:ver/custom/custom"              
#> [6,] "POST"     "/api/raw/cypher/transaction"           
#>      name                                                                    
#> [1,] "github.com/labstack/echo.glob..func1"                                  
#> [2,] "github.com/labstack/echo.glob..func1"                                  
#> [3,] "github.com/labstack/echo.glob..func1"                                  
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    
#> [6,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }