Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "FBl7"              "aL(L)"             "EPA(L)"           
#>   [4] "AL-DA2(R)"         "AL-VM4(R)"         "EPA(R)"           
#>   [7] "AL-VM3(R)"         "LX(R)"             "a'L(R)"           
#>  [10] "CRE(L)"            "ICL(L)"            "AL-VA3(R)"        
#>  [13] "VES(L)"            "AL-VM7v(R)"        "PB"               
#>  [16] "AL-VC5(R)"         "GA(R)"             "MB(L)"            
#>  [19] "CX"                "PENP"              "b2(R)"            
#>  [22] "BU(R)"             "SMP(R)"            "WED(R)"           
#>  [25] "FB"                "EBr1"              "SNP(R)"           
#>  [28] "AL-VL2a(R)"        "PB(R7)"            "AB(L)"            
#>  [31] "RUB(L)"            "AL-DA1(R)"         "AL-VC3l(R)"       
#>  [34] "AL-VP1d(R)"        "IB"                "EB"               
#>  [37] "AL-VM7d(L)"        "CRE(-RUB)(L)"      "NO2(R)"           
#>  [40] "AL-DM4(R)"         "SCL(L)"            "SPS(R)"           
#>  [43] "PB(L6)"            "b'L(L)"            "AL-DA4m(R)"       
#>  [46] "EBr3pw"            "AL-DP1l(R)"        "AL-VM5d(R)"       
#>  [49] "FBl5"              "AL-VP1l(R)"        "PB(L5)"           
#>  [52] "a'2(R)"            "AL-VM7v(L)"        "GOR(R)"           
#>  [55] "OL(R)"             "AL-D(R)"           "PB(L4)"           
#>  [58] "AL-VA1d(R)"        "AL-VP3(R)"         "LAL(-GA)(R)"      
#>  [61] "AL-DM6(L)"         "gL(R)"             "PB(L7)"           
#>  [64] "ICL(R)"            "AL-VA7m(R)"        "FBl6"             
#>  [67] "AL-DC2(R)"         "mALT(L)"           "AL-DC2(L)"        
#>  [70] "g5(R)"             "SAD(-AMMC)"        "FBl1"             
#>  [73] "a'L(L)"            "a2(R)"             "BU(L)"            
#>  [76] "NO(L)"             "AL-DM3(R)"         "EBr3am"           
#>  [79] "CA(L)"             "PB(R1)"            "AL-VA6(R)"        
#>  [82] "AL-VL1(R)"         "PB(L8)"            "AB(R)"            
#>  [85] "SLP(R)"            "AL-DL2d(R)"        "AL-VC2(R)"        
#>  [88] "EBr5"              "AL-DL5(L)"         "LO(R)"            
#>  [91] "MB(+ACA)(R)"       "AL-DA4l(R)"        "AL-DC3(R)"        
#>  [94] "AL-VP5(R)"         "VLNP(R)"           "AL-VC4(R)"        
#>  [97] "b'L(R)"            "AVLP(R)"           "LAL(R)"           
#> [100] "AL-VM5v(R)"        "AL-VM7d(R)"        "ROB(R)"           
#> [103] "FBl8"              "AL-DM2(R)"         "AL-DA3(L)"        
#> [106] "AL-DM4(L)"         "FBl4"              "AL-VP2(R)"        
#> [109] "AL-DA4m(L)"        "AL-DC4(L)"         "AL-VA7l(R)"       
#> [112] "AMMC"              "PRW"               "EBr3d"            
#> [115] "PB(R3)"            "b1(R)"             "PB(L1)"           
#> [118] "AL-DP1m(L)"        "IB112"             "PVLP(R)"          
#> [121] "b'2(R)"            "EBr6"              "g4(R)"            
#> [124] "PB(R4)"            "AL-VP1m(R)"        "RUB(R)"           
#> [127] "PB(L3)"            "dACA(R)"           "FLA(R)"           
#> [130] "FB-column3"        "AL-DA3(R)"         "SCL(R)"           
#> [133] "AL-DM5(L)"         "AL-DM3(L)"         "IPS(R)"           
#> [136] "a3(R)"             "AL-VA1v(R)"        "AL-VM2(R)"        
#> [139] "b'1(R)"            "CA(R)"             "AL-DL3(R)"        
#> [142] "AL-VC1(R)"         "POC"               "vACA(R)"          
#> [145] "PB(R6)"            "ME(R)"             "AL-D(L)"          
#> [148] "AL-DM5(R)"         "LOP(R)"            "bL(R)"            
#> [151] "PED(R)"            "a'3(R)"            "NO1(R)"           
#> [154] "NO3(L)"            "AME(R)"            "LX(L)"            
#> [157] "AL-DC1(R)"         "PB(L2)"            "g3(R)"            
#> [160] "aL(R)"             "g2(R)"             "GNG"              
#> [163] "AL-VA4(R)"         "EBr2r4"            "AL-VL2p(R)"       
#> [166] "GF(R)"             "NO3(R)"            "bL(L)"            
#> [169] "AL-VA5(R)"         "a1(R)"             "NO1(L)"           
#> [172] "PB(L9)"            "ATL(R)"            "CAN(R)"           
#> [175] "AOT(R)"            "CRE(-ROB,-RUB)(R)" "lACA(R)"          
#> [178] "PLP(R)"            "SNP(L)"            "AL-DC4(R)"        
#> [181] "AL-V(R)"           "LAL(L)"            "CRE(R)"           
#> [184] "SPS(L)"            "MB(R)"             "FBl9"             
#> [187] "g1(R)"             "PB(R5)"            "NO2(L)"           
#> [190] "GOR(L)"            "AL-DM1(R)"         "AL-DM6(R)"        
#> [193] "AL-VC3m(R)"        "AL-DA2(L)"         "AL-DC1(L)"        
#> [196] "FBl3"              "AL(R)"             "hemibrain"        
#> [199] "PB(R8)"            "AL-DM1(L)"         "AL-VA6(L)"        
#> [202] "VES(R)"            "AOTU(R)"           "PB(R2)"           
#> [205] "AL-DP1m(R)"        "GC"                "AL-DL4(L)"        
#> [208] "LH(R)"             "AL-DL5(R)"         "NO"               
#> [211] "INP"               "VMNP"              "AL-DM2(L)"        
#> [214] "SIP(R)"            "AL-DL4(R)"         "AL-VP4(R)"        
#> [217] "ATL(L)"            "SMP(L)"            "SAD"              
#> [220] "gL(L)"             "AL(L)"             "AL-VA2(R)"        
#> [223] "AL-VM1(R)"         "mALT(R)"           "PB(R9)"           
#> [226] "a'1(R)"            "FBl2"              "AL-DL2v(R)"       
#> [229] "SIP(L)"            "NO(R)"             "AL-DL1(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "b'1(R)"            "SIP(L)"            "SCL(R)"           
#>   [4] "g5(R)"             "PB(R3)"            "GNG"              
#>   [7] "dACA(R)"           "BU(L)"             "PVLP(R)"          
#>  [10] "FBl8"              "LAL(L)"            "AOTU(R)"          
#>  [13] "PB(R8)"            "PB(L6)"            "EPA(L)"           
#>  [16] "GA(R)"             "LX(L)"             "aL(L)"            
#>  [19] "GOR(R)"            "SCL(L)"            "LH(R)"            
#>  [22] "a'3(R)"            "CA(L)"             "b'2(R)"           
#>  [25] "EPA(R)"            "PB(R4)"            "ATL(L)"           
#>  [28] "IPS(R)"            "RUB(L)"            "bL(R)"            
#>  [31] "RUB(R)"            "vACA(R)"           "AME(R)"           
#>  [34] "AL(R)"             "NO2(L)"            "MB(L)"            
#>  [37] "PB(R1)"            "NO(R)"             "ICL(L)"           
#>  [40] "FBl5"              "NO"                "PB(L9)"           
#>  [43] "CRE(L)"            "VES(R)"            "PB(L5)"           
#>  [46] "PB(L3)"            "a'L(R)"            "SAD(-AMMC)"       
#>  [49] "a'L(L)"            "LX(R)"             "ME(R)"            
#>  [52] "BU(R)"             "AB(R)"             "a2(R)"            
#>  [55] "PLP(R)"            "LO(R)"             "ATL(R)"           
#>  [58] "VLNP(R)"           "OL(R)"             "FB"               
#>  [61] "FLA(R)"            "PB"                "VMNP"             
#>  [64] "EB"                "g2(R)"             "SPS(L)"           
#>  [67] "bL(L)"             "FBl1"              "b'L(L)"           
#>  [70] "gL(L)"             "SLP(R)"            "SPS(R)"           
#>  [73] "NO1(R)"            "CX"                "PENP"             
#>  [76] "PB(R6)"            "PRW"               "SMP(R)"           
#>  [79] "mALT(L)"           "g1(R)"             "MB(R)"            
#>  [82] "POC"               "EBr3am"            "VES(L)"           
#>  [85] "a1(R)"             "mALT(R)"           "gL(R)"            
#>  [88] "FBl6"              "NO2(R)"            "INP"              
#>  [91] "EBr1"              "PB(L7)"            "IB"               
#>  [94] "PB(L1)"            "EBr3d"             "AB(L)"            
#>  [97] "g3(R)"             "hemibrain"         "PED(R)"           
#> [100] "LAL(-GA)(R)"       "NO3(R)"            "GC"               
#> [103] "AMMC"              "SMP(L)"            "PB(R7)"           
#> [106] "GOR(L)"            "ROB(R)"            "PB(L2)"           
#> [109] "b'L(R)"            "EBr5"              "NO3(L)"           
#> [112] "AOT(R)"            "b1(R)"             "CAN(R)"           
#> [115] "lACA(R)"           "b2(R)"             "g4(R)"            
#> [118] "AL(L)"             "FBl7"              "FBl4"             
#> [121] "NO(L)"             "SAD"               "aL(R)"            
#> [124] "a'2(R)"            "MB(+ACA)(R)"       "FBl3"             
#> [127] "LOP(R)"            "CRE(R)"            "SNP(R)"           
#> [130] "PB(R9)"            "SNP(L)"            "FBl2"             
#> [133] "PB(R2)"            "a3(R)"             "PB(L4)"           
#> [136] "GF(R)"             "WED(R)"            "NO1(L)"           
#> [139] "SIP(R)"            "a'1(R)"            "CRE(-ROB,-RUB)(R)"
#> [142] "EBr6"              "EBr2r4"            "CA(R)"            
#> [145] "AVLP(R)"           "ICL(R)"            "PB(L8)"           
#> [148] "EBr3pw"            "FBl9"              "LAL(R)"           
#> [151] "PB(R5)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "PB(R8)"            "AL-DC4(L)"         "LO(R)"            
#>   [4] "FBl7"              "ATL(R)"            "AVLP(R)"          
#>   [7] "PVLP(R)"           "AL-DC2(R)"         "AL-VP1l(R)"       
#>  [10] "NO3(R)"            "CAN(R)"            "a3(R)"            
#>  [13] "AL-DM2(R)"         "CA(L)"             "SMP(L)"           
#>  [16] "AL-VA4(R)"         "AOT(R)"            "LOP(R)"           
#>  [19] "mALT(R)"           "PB(R3)"            "AL-DL2v(R)"       
#>  [22] "AL-VM7v(L)"        "PB(R1)"            "a'L(L)"           
#>  [25] "AL-VP1m(R)"        "GF(R)"             "AL-DM1(L)"        
#>  [28] "gL(R)"             "ICL(R)"            "AL-DA1(R)"        
#>  [31] "AL-VM2(R)"         "EBr3am"            "CX"               
#>  [34] "AB(L)"             "AL-DL5(L)"         "CA(R)"            
#>  [37] "AL-DL2d(R)"        "AL-DL5(R)"         "AL-VL2p(R)"       
#>  [40] "ME(R)"             "FB-column3"        "LAL(L)"           
#>  [43] "EBr1"              "LAL(R)"            "AL-DP1m(R)"       
#>  [46] "PB(R6)"            "FB"                "NO1(R)"           
#>  [49] "INP"               "AL-DL3(R)"         "PB(R7)"           
#>  [52] "SCL(R)"            "NO(R)"             "ICL(L)"           
#>  [55] "LX(L)"             "AMMC"              "ATL(L)"           
#>  [58] "PB(L9)"            "AL-DM6(R)"         "AL-VL2a(R)"       
#>  [61] "PB(R5)"            "SCL(L)"            "AL-VC5(R)"        
#>  [64] "AL-VM4(R)"         "EBr2r4"            "AL-DC2(L)"        
#>  [67] "FBl9"              "LH(R)"             "OL(R)"            
#>  [70] "GOR(R)"            "CRE(-ROB,-RUB)(R)" "g3(R)"            
#>  [73] "SAD"               "NO3(L)"            "FBl6"             
#>  [76] "IPS(R)"            "WED(R)"            "AL-DP1l(R)"       
#>  [79] "AL-VC2(R)"         "b2(R)"             "PB(L1)"           
#>  [82] "LX(R)"             "SNP(L)"            "AL-DL1(R)"        
#>  [85] "EPA(R)"            "AL-VA5(R)"         "AL-DC1(R)"        
#>  [88] "CRE(-RUB)(L)"      "b1(R)"             "MB(R)"            
#>  [91] "AL-VC3l(R)"        "VLNP(R)"           "GA(R)"            
#>  [94] "VMNP"              "AL-DC1(L)"         "b'L(L)"           
#>  [97] "PB(L5)"            "SIP(L)"            "lACA(R)"          
#> [100] "AL-DP1m(L)"        "NO"                "NO(L)"            
#> [103] "b'2(R)"            "FBl3"              "AL-VC1(R)"        
#> [106] "AL-VA1d(R)"        "AL-VM5v(R)"        "a'3(R)"           
#> [109] "AL-DM6(L)"         "AL-D(R)"           "AL-VA1v(R)"       
#> [112] "IB"                "GOR(L)"            "AL-DM3(L)"        
#> [115] "bL(R)"             "b'L(R)"            "EBr3pw"           
#> [118] "AL-VA7m(R)"        "SMP(R)"            "SNP(R)"           
#> [121] "VES(L)"            "AL-VM7v(R)"        "EBr6"             
#> [124] "a'L(R)"            "AL-DA3(L)"         "CRE(L)"           
#> [127] "FBl2"              "AL-VA3(R)"         "PRW"              
#> [130] "AL-DA4l(R)"        "NO1(L)"            "PB(L8)"           
#> [133] "RUB(L)"            "AL-VC3m(R)"        "AL-VP1d(R)"       
#> [136] "AB(R)"             "AL-DA3(R)"         "AL-VC4(R)"        
#> [139] "NO2(L)"            "AOTU(R)"           "g1(R)"            
#> [142] "MB(+ACA)(R)"       "mALT(L)"           "aL(R)"            
#> [145] "RUB(R)"            "AL-VP3(R)"         "AL-DA2(R)"        
#> [148] "AL-VL1(R)"         "AL-VM3(R)"         "PED(R)"           
#> [151] "EB"                "g4(R)"             "FBl5"             
#> [154] "AL-DL4(R)"         "AL-DM4(R)"         "dACA(R)"          
#> [157] "EBr3d"             "AL-VA6(L)"         "AL-DA4m(R)"       
#> [160] "AL-DC3(R)"         "PB(L3)"            "a'2(R)"           
#> [163] "EPA(L)"            "GNG"               "BU(L)"            
#> [166] "FLA(R)"            "AL-DA4m(L)"        "AL-DM5(R)"        
#> [169] "AME(R)"            "a2(R)"             "hemibrain"        
#> [172] "GC"                "PB(R9)"            "PB(L6)"           
#> [175] "bL(L)"             "PB(R2)"            "AL(R)"            
#> [178] "AL-VM7d(R)"        "b'1(R)"            "AL-DL4(L)"        
#> [181] "AL-DM2(L)"         "VES(R)"            "a1(R)"            
#> [184] "AL-DM4(L)"         "BU(R)"             "g5(R)"            
#> [187] "AL-V(R)"           "AL-VP5(R)"         "AL-DA2(L)"        
#> [190] "SLP(R)"            "AL-DC4(R)"         "AL-VP4(R)"        
#> [193] "MB(L)"             "EBr5"              "AL-D(L)"          
#> [196] "AL-DM5(L)"         "FBl4"              "PB"               
#> [199] "PB(R4)"            "AL-DM3(R)"         "AL-VM1(R)"        
#> [202] "PB(L7)"            "PB(L4)"            "AL-VP2(R)"        
#> [205] "SPS(L)"            "FBl8"              "AL-DM1(R)"        
#> [208] "AL-VA2(R)"         "PB(L2)"            "LAL(-GA)(R)"      
#> [211] "POC"               "CRE(R)"            "g2(R)"            
#> [214] "AL(L)"             "SAD(-AMMC)"        "AL-VA6(R)"        
#> [217] "AL-VM5d(R)"        "PENP"              "PLP(R)"           
#> [220] "AL-VM7d(L)"        "NO2(R)"            "aL(L)"            
#> [223] "SPS(R)"            "gL(L)"             "ROB(R)"           
#> [226] "FBl1"              "a'1(R)"            "vACA(R)"          
#> [229] "SIP(R)"            "AL-VA7l(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "a'L(R)"            "GA(R)"             "AL-DM2(L)"        
#>   [4] "mALT(L)"           "FBl1"              "LH(R)"            
#>   [7] "AL-VM2(R)"         "AL-DM5(L)"         "AL-VC3m(R)"       
#>  [10] "AL-VP1l(R)"        "LOP(R)"            "lACA(R)"          
#>  [13] "ICL(L)"            "FBl6"              "b'L(L)"           
#>  [16] "AL-VM5v(R)"        "GF(R)"             "VMNP"             
#>  [19] "SLP(R)"            "SMP(R)"            "SCL(L)"           
#>  [22] "LAL(R)"            "AL-DM4(R)"         "AL-DP1m(R)"       
#>  [25] "AL-VA1v(R)"        "AL-VA6(R)"         "FLA(R)"           
#>  [28] "EBr3pw"            "AL-VC3l(R)"        "ME(R)"            
#>  [31] "FBl4"              "IPS(R)"            "NO(L)"            
#>  [34] "a'L(L)"            "EBr3d"             "PB(L6)"           
#>  [37] "AL-VC1(R)"         "b2(R)"             "LO(R)"            
#>  [40] "b1(R)"             "OL(R)"             "AL-DM5(R)"        
#>  [43] "AL-VA7l(R)"        "b'2(R)"            "PB(L3)"           
#>  [46] "b'L(R)"            "PB(L9)"            "bL(R)"            
#>  [49] "AOT(R)"            "ATL(L)"            "PB(R8)"           
#>  [52] "PLP(R)"            "SIP(L)"            "EBr6"             
#>  [55] "AL-D(L)"           "CA(R)"             "MB(R)"            
#>  [58] "LAL(-GA)(R)"       "WED(R)"            "FB"               
#>  [61] "LX(L)"             "AL-VM4(R)"         "AMMC"             
#>  [64] "LX(R)"             "AL-DL2d(R)"        "AL-DL2v(R)"       
#>  [67] "PB(R5)"            "a'3(R)"            "GOR(L)"           
#>  [70] "AL-VP1m(R)"        "AL-VA6(L)"         "AL-VP1d(R)"       
#>  [73] "CA(L)"             "AOTU(R)"           "AL-DA2(L)"        
#>  [76] "NO3(L)"            "VES(R)"            "AL-VM1(R)"        
#>  [79] "AVLP(R)"           "PB(R4)"            "GOR(R)"           
#>  [82] "SPS(R)"            "PB(R9)"            "CRE(-RUB)(L)"     
#>  [85] "PVLP(R)"           "FBl7"              "AL-DL5(R)"        
#>  [88] "AL-VC4(R)"         "mALT(R)"           "AL(R)"            
#>  [91] "PRW"               "AB(R)"             "AL-DC3(R)"        
#>  [94] "AL-DL1(R)"         "AL-VC2(R)"         "CRE(R)"           
#>  [97] "AL-DA2(R)"         "AL-DM6(R)"         "AL-VL1(R)"        
#> [100] "AL-VP3(R)"         "AL-VA4(R)"         "g3(R)"            
#> [103] "AL-DA3(L)"         "AL-DA4m(L)"        "AL-DC4(L)"        
#> [106] "FBl5"              "PED(R)"            "PB"               
#> [109] "AL-DC2(L)"         "AL-VM7v(L)"        "AL-VA5(R)"        
#> [112] "AL-DM2(R)"         "LAL(L)"            "PB(R7)"           
#> [115] "PB(L1)"            "a'2(R)"            "RUB(L)"           
#> [118] "AL-V(R)"           "g2(R)"             "AL-DM4(L)"        
#> [121] "FB-column3"        "aL(L)"             "AL-DL4(R)"        
#> [124] "BU(R)"             "PB(L7)"            "AL-DC2(R)"        
#> [127] "NO1(R)"            "AL-DM1(L)"         "PB(L5)"           
#> [130] "NO"                "SAD(-AMMC)"        "FBl2"             
#> [133] "FBl9"              "SNP(L)"            "CX"               
#> [136] "AL-VM7d(R)"        "AL-VM7d(L)"        "CRE(L)"           
#> [139] "EPA(L)"            "AL-DA1(R)"         "AL-DL3(R)"        
#> [142] "AL-DM3(L)"         "AL-VP5(R)"         "b'1(R)"           
#> [145] "vACA(R)"           "EBr1"              "AL-VA7m(R)"       
#> [148] "CAN(R)"            "INP"               "a3(R)"            
#> [151] "AL-VL2a(R)"        "NO2(R)"            "AL-VP4(R)"        
#> [154] "g1(R)"             "g5(R)"             "AL-VL2p(R)"       
#> [157] "IB"                "MB(L)"             "g4(R)"            
#> [160] "AL-VC5(R)"         "SMP(L)"            "hemibrain"        
#> [163] "GC"                "aL(R)"             "AL-DC4(R)"        
#> [166] "SAD"               "AL-DP1m(L)"        "gL(R)"            
#> [169] "AL-DM3(R)"         "RUB(R)"            "NO(R)"            
#> [172] "ICL(R)"            "AL-VM3(R)"         "SCL(R)"           
#> [175] "SNP(R)"            "EBr2r4"            "gL(L)"            
#> [178] "AL-DL4(L)"         "PB(R3)"            "AL-DA3(R)"        
#> [181] "POC"               "EBr3am"            "PB(L4)"           
#> [184] "VES(L)"            "EBr5"              "PB(L8)"           
#> [187] "EB"                "PB(R1)"            "MB(+ACA)(R)"      
#> [190] "AL-DP1l(R)"        "AL(L)"             "AL-DM1(R)"        
#> [193] "PB(L2)"            "dACA(R)"           "AL-VA2(R)"        
#> [196] "AL-VA3(R)"         "PB(R6)"            "AL-DL5(L)"        
#> [199] "AL-DM6(L)"         "PB(R2)"            "AL-DC1(R)"        
#> [202] "a2(R)"             "AME(R)"            "AL-D(R)"          
#> [205] "a'1(R)"            "FBl3"              "FBl8"             
#> [208] "CRE(-ROB,-RUB)(R)" "NO2(L)"            "bL(L)"            
#> [211] "a1(R)"             "BU(L)"             "AL-DC1(L)"        
#> [214] "ATL(R)"            "AL-DA4l(R)"        "ROB(R)"           
#> [217] "GNG"               "PENP"              "NO3(R)"           
#> [220] "EPA(R)"            "SIP(R)"            "AL-VA1d(R)"       
#> [223] "NO1(L)"            "SPS(L)"            "AL-DA4m(R)"       
#> [226] "VLNP(R)"           "AB(L)"             "AL-VM5d(R)"       
#> [229] "AL-VM7v(R)"        "AL-VP2(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method  path                                        
#> [1,] "GET"   "/api/cached/roicompleteness"               
#> [2,] "GET"   "/api/dbmeta/database"                      
#> [3,] "POST"  "/api/v:ver/roimeshes/mesh/:dataset/:roi"   
#> [4,] "POST"  "/api/raw/cypher/cypher"                    
#> [5,] "POST"  "/api/v:ver/raw/cypher/transaction/:id/kill"
#> [6,] "TRACE" "/api/*"                                    
#>      name                                                                                    
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROICompleteness-fm"
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatabase-fm"        
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [6,] "github.com/labstack/echo.glob..func1"                                                  
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }