Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "SAD(-AMMC)"        "AL-DM6(R)"         "AL-VA5(R)"        
#>   [4] "AL-VA7l(R)"        "AL-VL2p(R)"        "AOT(R)"           
#>   [7] "EPA(R)"            "AL-VP1m(R)"        "AL-VP1l(R)"       
#>  [10] "CRE(-ROB,-RUB)(R)" "ATL(L)"            "PB(L6)"           
#>  [13] "AL-DM6(L)"         "AL-VM7d(L)"        "EPA(L)"           
#>  [16] "INP"               "GOR(R)"            "AL-DM5(R)"        
#>  [19] "PB(L1)"            "FBl4"              "AME(R)"           
#>  [22] "AL-VA6(L)"         "lACA(R)"           "NO3(R)"           
#>  [25] "b'L(R)"            "g4(R)"             "AL-DC4(L)"        
#>  [28] "FBl9"              "AL-VM5d(R)"        "PRW"              
#>  [31] "AL-VM7d(R)"        "bL(L)"             "AL-DA3(L)"        
#>  [34] "NO1(L)"            "FLA(R)"            "SNP(R)"           
#>  [37] "AL-DA2(R)"         "mALT(R)"           "AL-DM1(L)"        
#>  [40] "AL-DL3(R)"         "AL-DL5(R)"         "mALT(L)"          
#>  [43] "VMNP"              "PB(R1)"            "FB-column3"       
#>  [46] "ICL(R)"            "AL-VA3(R)"         "AL-VP3(R)"        
#>  [49] "PB(R5)"            "EBr3am"            "AOTU(R)"          
#>  [52] "FBl8"              "a'1(R)"            "ME(R)"            
#>  [55] "EBr3d"             "AL-DL5(L)"         "AL-VL2a(R)"       
#>  [58] "PB(L3)"            "MB(+ACA)(R)"       "LAL(L)"           
#>  [61] "BU(L)"             "VES(R)"            "g3(R)"            
#>  [64] "GNG"               "PB(L4)"            "AL-VC2(R)"        
#>  [67] "vACA(R)"           "a1(R)"             "PB(R7)"           
#>  [70] "AMMC"              "EBr5"              "a3(R)"            
#>  [73] "AL-VA4(R)"         "NO"                "PB(L7)"           
#>  [76] "IPS(R)"            "RUB(L)"            "AL-VP5(R)"        
#>  [79] "CA(L)"             "AL(L)"             "AL-VC5(R)"        
#>  [82] "AL-V(R)"           "AL-DL2v(R)"        "LX(R)"            
#>  [85] "AL-D(L)"           "AL-DL1(R)"         "CRE(R)"           
#>  [88] "LH(R)"             "AL-DL4(R)"         "PVLP(R)"          
#>  [91] "ICL(L)"            "AL-DA4m(R)"        "WED(R)"           
#>  [94] "gL(R)"             "g5(R)"             "AL-DC3(R)"        
#>  [97] "AL-DM4(L)"         "NO(L)"             "b1(R)"            
#> [100] "GC"                "SMP(L)"            "EBr1"             
#> [103] "gL(L)"             "FBl1"              "AL-VC1(R)"        
#> [106] "a'3(R)"            "GOR(L)"            "NO(R)"            
#> [109] "AL-DA4l(R)"        "AL-DM2(R)"         "b'1(R)"           
#> [112] "SPS(L)"            "bL(R)"             "AL-VA1v(R)"       
#> [115] "EB"                "IB112"             "AL-DM1(R)"        
#> [118] "AL-DM4(R)"         "AL-VM7v(R)"        "RUB(R)"           
#> [121] "a2(R)"             "BU(R)"             "LX(L)"            
#> [124] "AL-DA3(R)"         "LOP(R)"            "PED(R)"           
#> [127] "LO(R)"             "AVLP(R)"           "AL-DC2(R)"        
#> [130] "AL-DP1l(R)"        "MB(L)"             "AL(R)"            
#> [133] "PB(R8)"            "AL-DC1(L)"         "AL-DM3(L)"        
#> [136] "AL-DP1m(L)"        "FBl6"              "AL-D(R)"          
#> [139] "PENP"              "SMP(R)"            "PB(L5)"           
#> [142] "SCL(R)"            "EBr3pw"            "OL(R)"            
#> [145] "AL-DC4(R)"         "a'L(R)"            "AL-VM7v(L)"       
#> [148] "PB(L9)"            "hemibrain"         "NO1(R)"           
#> [151] "PB"                "AL-VA1d(R)"        "AL-VL1(R)"        
#> [154] "POC"               "LAL(-GA)(R)"       "AL-VM1(R)"        
#> [157] "VLNP(R)"           "aL(R)"             "MB(R)"            
#> [160] "AL-VP1d(R)"        "NO2(R)"            "AL-VM4(R)"        
#> [163] "SCL(L)"            "AL-DL4(L)"         "g2(R)"            
#> [166] "CA(R)"             "SIP(L)"            "AL-DC1(R)"        
#> [169] "a'2(R)"            "CRE(L)"            "NO3(L)"           
#> [172] "LAL(R)"            "PLP(R)"            "AL-DM2(L)"        
#> [175] "AL-VC3l(R)"        "FBl5"              "FBl7"             
#> [178] "AL-DC2(L)"         "b'2(R)"            "FB"               
#> [181] "EBr6"              "AL-DA2(L)"         "AL-DA4m(L)"       
#> [184] "PB(R6)"            "PB(L8)"            "ATL(R)"           
#> [187] "AL-DL2d(R)"        "AL-VC3m(R)"        "GF(R)"            
#> [190] "IB"                "VES(L)"            "b'L(L)"           
#> [193] "g1(R)"             "b2(R)"             "PB(R9)"           
#> [196] "SPS(R)"            "PB(R2)"            "AL-DP1m(R)"       
#> [199] "AL-VA2(R)"         "AL-VA6(R)"         "CX"               
#> [202] "SIP(R)"            "FBl2"              "SAD"              
#> [205] "aL(L)"             "AL-VA7m(R)"        "EBr2r4"           
#> [208] "ROB(R)"            "PB(R4)"            "AL-DM3(R)"        
#> [211] "AL-VC4(R)"         "dACA(R)"           "AL-VM3(R)"        
#> [214] "AL-VP4(R)"         "SNP(L)"            "AB(R)"            
#> [217] "AL-VM5v(R)"        "PB(R3)"            "NO2(L)"           
#> [220] "a'L(L)"            "CRE(-RUB)(L)"      "CAN(R)"           
#> [223] "SLP(R)"            "AL-DA1(R)"         "PB(L2)"           
#> [226] "GA(R)"             "FBl3"              "AL-VM2(R)"        
#> [229] "AL-VP2(R)"         "AB(L)"             "AL-DM5(L)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "BU(L)"             "bL(L)"             "PB(R2)"           
#>   [4] "AVLP(R)"           "PB(L3)"            "PRW"              
#>   [7] "mALT(R)"           "WED(R)"            "NO2(R)"           
#>  [10] "POC"               "b'1(R)"            "PED(R)"           
#>  [13] "aL(R)"             "PVLP(R)"           "MB(+ACA)(R)"      
#>  [16] "LAL(L)"            "RUB(L)"            "PB(R5)"           
#>  [19] "PB(L4)"            "AL(R)"             "SMP(L)"           
#>  [22] "g2(R)"             "ROB(R)"            "a3(R)"            
#>  [25] "PENP"              "PLP(R)"            "AB(R)"            
#>  [28] "b1(R)"             "EPA(L)"            "OL(R)"            
#>  [31] "SIP(L)"            "g1(R)"             "FLA(R)"           
#>  [34] "FBl4"              "SLP(R)"            "IB"               
#>  [37] "dACA(R)"           "PB(R1)"            "FBl5"             
#>  [40] "PB(R4)"            "LAL(R)"            "PB(L8)"           
#>  [43] "AL(L)"             "EBr3pw"            "FBl6"             
#>  [46] "SCL(R)"            "a'3(R)"            "hemibrain"        
#>  [49] "gL(R)"             "AOT(R)"            "PB(L5)"           
#>  [52] "SPS(R)"            "LAL(-GA)(R)"       "ATL(R)"           
#>  [55] "AB(L)"             "AME(R)"            "b'2(R)"           
#>  [58] "AOTU(R)"           "EBr5"              "EB"               
#>  [61] "LO(R)"             "IPS(R)"            "GC"               
#>  [64] "FB"                "EBr6"              "ICL(L)"           
#>  [67] "CA(R)"             "CAN(R)"            "INP"              
#>  [70] "PB(L2)"            "GF(R)"             "GOR(L)"           
#>  [73] "PB"                "BU(R)"             "PB(L9)"           
#>  [76] "SAD(-AMMC)"        "LX(L)"             "bL(R)"            
#>  [79] "SAD"               "EBr1"              "PB(L1)"           
#>  [82] "ME(R)"             "MB(R)"             "EBr3am"           
#>  [85] "FBl7"              "FBl8"              "GOR(R)"           
#>  [88] "CA(L)"             "PB(R7)"            "NO3(R)"           
#>  [91] "CRE(R)"            "EBr2r4"            "VLNP(R)"          
#>  [94] "mALT(L)"           "SPS(L)"            "g4(R)"            
#>  [97] "EPA(R)"            "ICL(R)"            "SNP(R)"           
#> [100] "LH(R)"             "ATL(L)"            "PB(L6)"           
#> [103] "FBl1"              "SNP(L)"            "FBl9"             
#> [106] "GNG"               "AMMC"              "FBl3"             
#> [109] "a'L(R)"            "NO3(L)"            "PB(L7)"           
#> [112] "vACA(R)"           "g3(R)"             "lACA(R)"          
#> [115] "b2(R)"             "NO1(R)"            "CX"               
#> [118] "a2(R)"             "gL(L)"             "VES(L)"           
#> [121] "RUB(R)"            "CRE(L)"            "FBl2"             
#> [124] "LX(R)"             "EBr3d"             "NO"               
#> [127] "NO(R)"             "PB(R6)"            "PB(R9)"           
#> [130] "NO2(L)"            "NO1(L)"            "g5(R)"            
#> [133] "PB(R3)"            "b'L(R)"            "LOP(R)"           
#> [136] "SIP(R)"            "aL(L)"             "b'L(L)"           
#> [139] "NO(L)"             "CRE(-ROB,-RUB)(R)" "GA(R)"            
#> [142] "PB(R8)"            "VMNP"              "a'2(R)"           
#> [145] "SMP(R)"            "SCL(L)"            "a1(R)"            
#> [148] "MB(L)"             "a'1(R)"            "VES(R)"           
#> [151] "a'L(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "NO3(R)"            "AL-D(L)"           "vACA(R)"          
#>   [4] "AL-VL1(R)"         "SMP(L)"            "a'2(R)"           
#>   [7] "VES(R)"            "GNG"               "b'L(L)"           
#>  [10] "AL-VC2(R)"         "b2(R)"             "AL-DC1(L)"        
#>  [13] "PB(L6)"            "PB(R2)"            "NO(L)"            
#>  [16] "AL-DA3(R)"         "mALT(L)"           "WED(R)"           
#>  [19] "AL-DC1(R)"         "AL-VM5d(R)"        "SCL(R)"           
#>  [22] "AL-VA6(R)"         "a2(R)"             "FBl8"             
#>  [25] "AVLP(R)"           "AL-DM5(R)"         "PED(R)"           
#>  [28] "AL-DM4(L)"         "AL-DC2(R)"         "AL-VM5v(R)"       
#>  [31] "AOT(R)"            "PB(L5)"            "LAL(L)"           
#>  [34] "g1(R)"             "LX(R)"             "AL-VM2(R)"        
#>  [37] "SCL(L)"            "CRE(R)"            "SPS(L)"           
#>  [40] "aL(L)"             "AL-VA1v(R)"        "POC"              
#>  [43] "NO(R)"             "EBr3pw"            "INP"              
#>  [46] "AL-VP1d(R)"        "CX"                "GF(R)"            
#>  [49] "AL-DA2(L)"         "ICL(L)"            "FBl2"             
#>  [52] "FBl4"              "GC"                "g4(R)"            
#>  [55] "RUB(R)"            "mALT(R)"           "LO(R)"            
#>  [58] "FB-column3"        "CA(R)"             "AL-DM1(R)"        
#>  [61] "AL-D(R)"           "AL-VM1(R)"         "ROB(R)"           
#>  [64] "FBl9"              "PB(R5)"            "PB(R4)"           
#>  [67] "LOP(R)"            "aL(R)"             "SIP(R)"           
#>  [70] "PB(L9)"            "PB(L2)"            "SIP(L)"           
#>  [73] "FBl3"              "AL-VA7l(R)"        "ATL(L)"           
#>  [76] "g2(R)"             "PB(L1)"            "AL-DM5(L)"        
#>  [79] "AL-DP1m(L)"        "CRE(-RUB)(L)"      "SMP(R)"           
#>  [82] "PENP"              "SAD"               "SNP(R)"           
#>  [85] "gL(L)"             "AL-VM7v(L)"        "SNP(L)"           
#>  [88] "SLP(R)"            "AL-DC2(L)"         "ICL(R)"           
#>  [91] "EBr5"              "NO1(R)"            "EB"               
#>  [94] "AL-DM6(L)"         "FBl6"              "AL-DM6(R)"        
#>  [97] "MB(L)"             "AL-DM3(L)"         "AL-DA4m(R)"       
#> [100] "AL-VC3l(R)"        "VLNP(R)"           "b'L(R)"           
#> [103] "NO"                "PVLP(R)"           "AB(R)"            
#> [106] "LAL(R)"            "EBr1"              "AL-VL2a(R)"       
#> [109] "AL-VP1l(R)"        "g3(R)"             "GOR(L)"           
#> [112] "AL-DM1(L)"         "bL(R)"             "AL-VA5(R)"        
#> [115] "AL-DL4(L)"         "AL-DP1m(R)"        "AL-VM4(R)"        
#> [118] "AL-VM7d(L)"        "NO2(R)"            "CAN(R)"           
#> [121] "AL-VA3(R)"         "NO2(L)"            "NO3(L)"           
#> [124] "AL-VC4(R)"         "AL-VL2p(R)"        "a1(R)"            
#> [127] "AL-DL2d(R)"        "AMMC"              "a'L(R)"           
#> [130] "MB(+ACA)(R)"       "AL-DC4(R)"         "PRW"              
#> [133] "AL-VA4(R)"         "LAL(-GA)(R)"       "PLP(R)"           
#> [136] "ME(R)"             "LH(R)"             "VES(L)"           
#> [139] "MB(R)"             "AL-VA2(R)"         "AL-DC4(L)"        
#> [142] "PB(R9)"            "EPA(L)"            "PB(L3)"           
#> [145] "FLA(R)"            "PB(R1)"            "PB(R3)"           
#> [148] "AL-DL1(R)"         "AL-VP4(R)"         "EBr6"             
#> [151] "VMNP"              "AL-DP1l(R)"        "AL-VP3(R)"        
#> [154] "ATL(R)"            "AL-DA4m(L)"        "CRE(-ROB,-RUB)(R)"
#> [157] "EBr3d"             "PB(R6)"            "AL-DM4(R)"        
#> [160] "AL-VC1(R)"         "AL-VM7v(R)"        "AL-V(R)"          
#> [163] "AL-VP5(R)"         "IB"                "BU(L)"            
#> [166] "FB"                "gL(R)"             "AL-VC3m(R)"       
#> [169] "BU(R)"             "a3(R)"             "AOTU(R)"          
#> [172] "FBl1"              "IPS(R)"            "OL(R)"            
#> [175] "AL-DC3(R)"         "hemibrain"         "PB(R8)"           
#> [178] "AL-VP2(R)"         "a'3(R)"            "CRE(L)"           
#> [181] "FBl7"              "RUB(L)"            "AL-DA1(R)"        
#> [184] "AL-VC5(R)"         "a'1(R)"            "SAD(-AMMC)"       
#> [187] "FBl5"              "AL-DA2(R)"         "AL-DL2v(R)"       
#> [190] "AB(L)"             "PB"                "AL-DA3(L)"        
#> [193] "LX(L)"             "AL-VM3(R)"         "NO1(L)"           
#> [196] "AL-DL5(L)"         "g5(R)"             "AL-DA4l(R)"       
#> [199] "AL-VA1d(R)"        "AL-VA7m(R)"        "bL(L)"            
#> [202] "b1(R)"             "b'1(R)"            "AL(L)"            
#> [205] "AL-DM2(L)"         "AL-VA6(L)"         "GOR(R)"           
#> [208] "AL-VM7d(R)"        "SPS(R)"            "AL-DL5(R)"        
#> [211] "AL-DM2(R)"         "AL-VP1m(R)"        "AME(R)"           
#> [214] "AL-DL3(R)"         "dACA(R)"           "EBr2r4"           
#> [217] "AL-DL4(R)"         "EBr3am"            "b'2(R)"           
#> [220] "PB(R7)"            "EPA(R)"            "PB(L7)"           
#> [223] "PB(L4)"            "a'L(L)"            "AL-DM3(R)"        
#> [226] "CA(L)"             "GA(R)"             "lACA(R)"          
#> [229] "PB(L8)"            "AL(R)"            
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.2`
#> $`hemibrain:v1.2`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2`$ROIs
#>   [1] "b'2(R)"            "EBr1"              "AL-DM6(L)"        
#>   [4] "PB(L6)"            "PB(L1)"            "PB(R8)"           
#>   [7] "LAL(R)"            "AL-VL1(R)"         "GNG"              
#>  [10] "AL-DC2(R)"         "POC"               "AL-DM3(L)"        
#>  [13] "EBr3pw"            "EPA(L)"            "SMP(R)"           
#>  [16] "AL-DA4m(L)"        "PB(R2)"            "AL-VC2(R)"        
#>  [19] "NO1(L)"            "EBr2r4"            "AL-DL4(L)"        
#>  [22] "PB(L7)"            "FBl3"              "AL-DA4m(R)"       
#>  [25] "AL-DL2d(R)"        "GC"                "MB(+ACA)(R)"      
#>  [28] "AL-VA6(R)"         "AL-DA3(L)"         "PB(R1)"           
#>  [31] "FBl6"              "AL-DM5(R)"         "AL-VM2(R)"        
#>  [34] "BU(R)"             "AL-DM3(R)"         "AL-DP1l(R)"       
#>  [37] "AL-VA5(R)"         "AL-VM7v(R)"        "g3(R)"            
#>  [40] "LAL(-GA)(R)"       "g2(R)"             "NO1(R)"           
#>  [43] "SAD(-AMMC)"        "AL(R)"             "AL-VM3(R)"        
#>  [46] "AOT(R)"            "AME(R)"            "GOR(R)"           
#>  [49] "AL-VL2a(R)"        "EBr6"              "BU(L)"            
#>  [52] "AL-DM2(L)"         "AL-DP1m(L)"        "ICL(L)"           
#>  [55] "PB(L5)"            "NO3(L)"            "AL-VA7l(R)"       
#>  [58] "FB"                "PB(L8)"            "gL(L)"            
#>  [61] "PB(L9)"            "AL-VM5d(R)"        "g1(R)"            
#>  [64] "PRW"               "EBr3d"             "AL-DC3(R)"        
#>  [67] "AL-V(R)"           "VMNP"              "ICL(R)"           
#>  [70] "IB"                "GA(R)"             "SAD"              
#>  [73] "PB"                "AL-DC1(L)"         "gL(R)"            
#>  [76] "PB(L2)"            "AMMC"              "AL-VA7m(R)"       
#>  [79] "PENP"              "GF(R)"             "CRE(R)"           
#>  [82] "FB-column3"        "PB(L4)"            "SLP(R)"           
#>  [85] "AL-DL5(R)"         "CRE(-RUB)(L)"      "g5(R)"            
#>  [88] "FBl5"              "AL-DA2(R)"         "AL-VM1(R)"        
#>  [91] "AL-VP1d(R)"        "b'1(R)"            "SNP(R)"           
#>  [94] "OL(R)"             "AL-DA1(R)"         "EBr3am"           
#>  [97] "FBl7"              "AL-VP1l(R)"        "LX(L)"            
#> [100] "AL-DL4(R)"         "AL-VL2p(R)"        "CRE(-ROB,-RUB)(R)"
#> [103] "a2(R)"             "AL-DC1(R)"         "AL-DL3(R)"        
#> [106] "AL-VP2(R)"         "NO2(L)"            "bL(L)"            
#> [109] "FBl4"              "AL-DL1(R)"         "vACA(R)"          
#> [112] "SIP(R)"            "AOTU(R)"           "mALT(R)"          
#> [115] "FBl9"              "a'3(R)"            "hemibrain"        
#> [118] "lACA(R)"           "a'1(R)"            "LO(R)"            
#> [121] "VES(R)"            "AL-VA1v(R)"        "AL-VP4(R)"        
#> [124] "RUB(R)"            "LAL(L)"            "g4(R)"            
#> [127] "NO"                "EB"                "ATL(R)"           
#> [130] "PB(L3)"            "AL-VM7v(L)"        "RUB(L)"           
#> [133] "WED(R)"            "AL-VA6(L)"         "IPS(R)"           
#> [136] "mALT(L)"           "ROB(R)"            "FBl2"             
#> [139] "PB(R5)"            "CA(L)"             "EBr5"             
#> [142] "PB(R4)"            "AL-DM2(R)"         "AL-VM7d(R)"       
#> [145] "MB(L)"             "b'L(R)"            "AL-VC4(R)"        
#> [148] "AL-VC5(R)"         "a1(R)"             "SCL(R)"           
#> [151] "SPS(L)"            "ME(R)"             "SNP(L)"           
#> [154] "aL(L)"             "PB(R6)"            "AL-DM1(L)"        
#> [157] "LOP(R)"            "SIP(L)"            "SPS(R)"           
#> [160] "AL-DA4l(R)"        "AL(L)"             "FBl8"             
#> [163] "PB(R3)"            "AL-DC4(R)"         "AL-DP1m(R)"       
#> [166] "AL-VP5(R)"         "SMP(L)"            "NO(R)"            
#> [169] "NO(L)"             "AL-DA3(R)"         "AL-DM4(R)"        
#> [172] "AL-DM6(R)"         "AL-D(R)"           "AL-DL5(L)"        
#> [175] "a3(R)"             "AL-DM1(R)"         "AL-VC1(R)"        
#> [178] "CRE(L)"            "AL-VA2(R)"         "AL-VA3(R)"        
#> [181] "AL-VC3l(R)"        "PED(R)"            "PLP(R)"           
#> [184] "AL-DC2(L)"         "AL-VC3m(R)"        "a'L(R)"           
#> [187] "SCL(L)"            "AL-VM7d(L)"        "CA(R)"            
#> [190] "AB(R)"             "VLNP(R)"           "b2(R)"            
#> [193] "LH(R)"             "AL-VA1d(R)"        "AL-VA4(R)"        
#> [196] "CX"                "PB(R7)"            "NO2(R)"           
#> [199] "AL-VM4(R)"         "AL-DA2(L)"         "AL-DC4(L)"        
#> [202] "PB(R9)"            "CAN(R)"            "FLA(R)"           
#> [205] "b1(R)"             "AL-DL2v(R)"        "AL-VM5v(R)"       
#> [208] "AL-VP3(R)"         "AL-DM4(L)"         "AVLP(R)"          
#> [211] "bL(R)"             "b'L(L)"            "VES(L)"           
#> [214] "AL-DM5(L)"         "GOR(L)"            "LX(R)"            
#> [217] "a'2(R)"            "EPA(R)"            "INP"              
#> [220] "ATL(L)"            "AL-D(L)"           "PVLP(R)"          
#> [223] "FBl1"              "a'L(L)"            "aL(R)"            
#> [226] "NO3(R)"            "MB(R)"             "AL-VP1m(R)"       
#> [229] "dACA(R)"           "AB(L)"            
#> 
#> $`hemibrain:v1.2`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method     path                                
#> [1,] "POST"     "/api/v:ver/npexplorer/distribution"
#> [2,] "GET"      "/api/v:ver/cached/roiconnectivity" 
#> [3,] "POST"     "/api/v:ver/npexplorer/roisinneuron"
#> [4,] "GET"      "/api/npexplorer/findneurons"       
#> [5,] "HEAD"     "/api/*"                            
#> [6,] "PROPFIND" "/api/*"                            
#>      name                                                                                       
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getFindNeurons-fm"
#> [5,] "github.com/labstack/echo.glob..func1"                                                     
#> [6,] "github.com/labstack/echo.glob..func1"                                                     
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }