Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DA4m(L)"        "AL-DC2(L)"         "AL-D(L)"          
#>   [4] "AL-VA6(L)"         "a'3(R)"            "a2(R)"            
#>   [7] "g1(R)"             "EBr5"              "AL-DC4(R)"        
#>  [10] "mALT(L)"           "PED(R)"            "CAN(R)"           
#>  [13] "AL-DC3(R)"         "EB"                "AL-V(R)"          
#>  [16] "NO3(R)"            "LH(R)"             "AL-VC4(R)"        
#>  [19] "SNP(L)"            "NO2(L)"            "FBl8"             
#>  [22] "CA(R)"             "AL-DA1(R)"         "AL-DL4(R)"        
#>  [25] "AL-DC1(L)"         "AL-VM7d(L)"        "GNG"              
#>  [28] "NO2(R)"            "AL-VC2(R)"         "AL-DM6(R)"        
#>  [31] "AL-VP1l(R)"        "PB(R9)"            "b'1(R)"           
#>  [34] "FBl7"              "PB(R2)"            "AL-VL1(R)"        
#>  [37] "SAD"               "PB"                "g4(R)"            
#>  [40] "LX(R)"             "a'2(R)"            "SAD(-AMMC)"       
#>  [43] "AB(R)"             "GC"                "AL-VP4(R)"        
#>  [46] "gL(L)"             "AVLP(R)"           "hemibrain"        
#>  [49] "AL-VM3(R)"         "IB112"             "FBl9"             
#>  [52] "mALT(R)"           "BU(L)"             "PB(R1)"           
#>  [55] "gL(R)"             "PB(L5)"            "AL-DL3(R)"        
#>  [58] "FBl3"              "FBl6"              "AME(R)"           
#>  [61] "SNP(R)"            "AL(R)"             "AL-VA3(R)"        
#>  [64] "EBr1"              "SCL(L)"            "NO1(R)"           
#>  [67] "ATL(R)"            "AL-DL2v(R)"        "FB"               
#>  [70] "AL-DM6(L)"         "AL-DL5(R)"         "CX"               
#>  [73] "MB(R)"             "SMP(R)"            "GF(R)"            
#>  [76] "lACA(R)"           "FBl5"              "AL-VC3m(R)"       
#>  [79] "AL-DM3(R)"         "IPS(R)"            "bL(R)"            
#>  [82] "a'L(L)"            "SPS(L)"            "CRE(-RUB)(L)"     
#>  [85] "OL(R)"             "AL-DP1m(R)"        "AL-VA6(R)"        
#>  [88] "AL-VM5d(R)"        "SIP(R)"            "ROB(R)"           
#>  [91] "PB(L6)"            "CA(L)"             "CRE(L)"           
#>  [94] "AL-VM5v(R)"        "AL-VP2(R)"         "AL-DL2d(R)"       
#>  [97] "AL-VM4(R)"         "AL-VP3(R)"         "EBr3am"           
#> [100] "b2(R)"             "ME(R)"             "EBr2r4"           
#> [103] "VES(R)"            "AL-DM4(R)"         "PRW"              
#> [106] "FLA(R)"            "g5(R)"             "AL-DM1(R)"        
#> [109] "NO3(L)"            "ICL(L)"            "AL-VA2(R)"        
#> [112] "aL(R)"             "PB(R8)"            "NO(R)"            
#> [115] "SLP(R)"            "AOT(R)"            "AB(L)"            
#> [118] "RUB(R)"            "AL-DM5(L)"         "NO"               
#> [121] "GOR(R)"            "AL-DM2(R)"         "AL-DP1l(R)"       
#> [124] "AL-D(R)"           "AL-VL2a(R)"        "EPA(R)"           
#> [127] "PB(L7)"            "AL-DA4m(R)"        "PB(R3)"           
#> [130] "AL-VL2p(R)"        "VLNP(R)"           "EBr6"             
#> [133] "AL-VA7m(R)"        "aL(L)"             "PB(R4)"           
#> [136] "NO(L)"             "AL-VM2(R)"         "AL-VP1m(R)"       
#> [139] "SPS(R)"            "MB(L)"             "EBr3pw"           
#> [142] "AL-VP5(R)"         "CRE(R)"            "AL(L)"            
#> [145] "AL-DP1m(L)"        "AL-VA7l(R)"        "PB(L2)"           
#> [148] "vACA(R)"           "FBl4"              "ICL(R)"           
#> [151] "VES(L)"            "FB-column3"        "PB(R6)"           
#> [154] "SMP(L)"            "AL-DM3(L)"         "bL(L)"            
#> [157] "b1(R)"             "SCL(R)"            "EBr3d"            
#> [160] "AL-DA3(L)"         "AL-DA2(R)"         "AL-VM7v(R)"       
#> [163] "PB(R5)"            "WED(R)"            "PB(L1)"           
#> [166] "VMNP"              "AL-VM7v(L)"        "PB(R7)"           
#> [169] "GOR(L)"            "PLP(R)"            "PVLP(R)"          
#> [172] "AL-VP1d(R)"        "AOTU(R)"           "PB(L9)"           
#> [175] "AL-DM5(R)"         "INP"               "AL-DA2(L)"        
#> [178] "AL-DL4(L)"         "FBl1"              "AL-DA4l(R)"       
#> [181] "AL-VC1(R)"         "AL-VM1(R)"         "AL-VM7d(R)"       
#> [184] "POC"               "PENP"              "LAL(L)"           
#> [187] "PB(L4)"            "a'1(R)"            "MB(+ACA)(R)"      
#> [190] "FBl2"              "LX(L)"             "AL-DC2(R)"        
#> [193] "g3(R)"             "GA(R)"             "g2(R)"            
#> [196] "AL-DL1(R)"         "NO1(L)"            "dACA(R)"          
#> [199] "AL-DC1(R)"         "a'L(R)"            "AL-DA3(R)"        
#> [202] "a1(R)"             "SIP(L)"            "BU(R)"            
#> [205] "b'L(L)"            "IB"                "AL-DM1(L)"        
#> [208] "EPA(L)"            "RUB(L)"            "PB(L8)"           
#> [211] "AL-VA1d(R)"        "PB(L3)"            "LOP(R)"           
#> [214] "AMMC"              "LAL(-GA)(R)"       "AL-DC4(L)"        
#> [217] "AL-VC3l(R)"        "ATL(L)"            "a3(R)"            
#> [220] "AL-VA1v(R)"        "AL-VC5(R)"         "b'L(R)"           
#> [223] "LAL(R)"            "CRE(-ROB,-RUB)(R)" "AL-DL5(L)"        
#> [226] "LO(R)"             "AL-VA5(R)"         "AL-VA4(R)"        
#> [229] "b'2(R)"            "AL-DM2(L)"         "AL-DM4(L)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "EB"                "EPA(R)"            "a'1(R)"           
#>   [4] "NO3(L)"            "MB(R)"             "PB(R5)"           
#>   [7] "LAL(L)"            "SIP(L)"            "PVLP(R)"          
#>  [10] "FBl9"              "bL(R)"             "NO"               
#>  [13] "NO(R)"             "SMP(L)"            "g2(R)"            
#>  [16] "SCL(R)"            "SIP(R)"            "LO(R)"            
#>  [19] "FBl8"              "mALT(R)"           "NO3(R)"           
#>  [22] "EBr5"              "VES(R)"            "MB(+ACA)(R)"      
#>  [25] "FBl7"              "EPA(L)"            "b'1(R)"           
#>  [28] "g3(R)"             "GOR(L)"            "AB(L)"            
#>  [31] "gL(L)"             "SAD(-AMMC)"        "FB"               
#>  [34] "SPS(L)"            "ROB(R)"            "PB(L6)"           
#>  [37] "g5(R)"             "GNG"               "PED(R)"           
#>  [40] "b2(R)"             "IB"                "ATL(L)"           
#>  [43] "LAL(R)"            "PB(L4)"            "a'3(R)"           
#>  [46] "CRE(R)"            "b1(R)"             "a'L(L)"           
#>  [49] "CA(L)"             "PB(R7)"            "INP"              
#>  [52] "AB(R)"             "AME(R)"            "GOR(R)"           
#>  [55] "b'L(L)"            "EBr1"              "AL(L)"            
#>  [58] "PB(L8)"            "EBr3pw"            "FBl5"             
#>  [61] "ATL(R)"            "VLNP(R)"           "PRW"              
#>  [64] "SPS(R)"            "LH(R)"             "IPS(R)"           
#>  [67] "mALT(L)"           "VMNP"              "a'2(R)"           
#>  [70] "bL(L)"             "SLP(R)"            "POC"              
#>  [73] "FLA(R)"            "EBr3d"             "EBr6"             
#>  [76] "FBl2"              "ICL(R)"            "WED(R)"           
#>  [79] "NO1(L)"            "CRE(L)"            "VES(L)"           
#>  [82] "lACA(R)"           "LAL(-GA)(R)"       "PB(R2)"           
#>  [85] "GA(R)"             "PB(R3)"            "RUB(L)"           
#>  [88] "PB(L3)"            "aL(R)"             "gL(R)"            
#>  [91] "FBl4"              "EBr2r4"            "a'L(R)"           
#>  [94] "FBl6"              "PB"                "PB(L7)"           
#>  [97] "b'L(R)"            "LX(L)"             "PENP"             
#> [100] "SAD"               "g1(R)"             "MB(L)"            
#> [103] "b'2(R)"            "GF(R)"             "CAN(R)"           
#> [106] "PB(R4)"            "NO(L)"             "FBl1"             
#> [109] "NO2(R)"            "aL(L)"             "RUB(R)"           
#> [112] "vACA(R)"           "g4(R)"             "AVLP(R)"          
#> [115] "PLP(R)"            "LX(R)"             "SNP(R)"           
#> [118] "BU(R)"             "SNP(L)"            "PB(L5)"           
#> [121] "a2(R)"             "dACA(R)"           "a1(R)"            
#> [124] "PB(L9)"            "ICL(L)"            "a3(R)"            
#> [127] "AMMC"              "PB(R9)"            "PB(R1)"           
#> [130] "EBr3am"            "PB(L1)"            "SMP(R)"           
#> [133] "hemibrain"         "ME(R)"             "PB(L2)"           
#> [136] "CX"                "OL(R)"             "AL(R)"            
#> [139] "PB(R8)"            "FBl3"              "CA(R)"            
#> [142] "SCL(L)"            "LOP(R)"            "CRE(-ROB,-RUB)(R)"
#> [145] "GC"                "NO1(R)"            "NO2(L)"           
#> [148] "BU(L)"             "AOT(R)"            "PB(R6)"           
#> [151] "AOTU(R)"          
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "a'1(R)"            "ATL(L)"            "AMMC"             
#>   [4] "dACA(R)"           "PB(L8)"            "PB(L6)"           
#>   [7] "PVLP(R)"           "PB(R3)"            "b'2(R)"           
#>  [10] "aL(R)"             "ATL(R)"            "AL-DL2v(R)"       
#>  [13] "EBr5"              "PLP(R)"            "AL-VM5v(R)"       
#>  [16] "AL-DM6(L)"         "AL-DM4(R)"         "a'L(R)"           
#>  [19] "g4(R)"             "AL-DC4(L)"         "AL-DL5(L)"        
#>  [22] "AL-VM7d(L)"        "AL(R)"             "LAL(L)"           
#>  [25] "WED(R)"            "CA(R)"             "AL-VL2p(R)"       
#>  [28] "ICL(L)"            "SIP(L)"            "PB(R1)"           
#>  [31] "NO3(R)"            "AL-DA4m(L)"        "AL-DA3(L)"        
#>  [34] "AL-DC2(L)"         "AL-DM4(L)"         "PB(R9)"           
#>  [37] "LAL(R)"            "AL-DP1m(R)"        "lACA(R)"          
#>  [40] "EBr6"              "AL-DA2(R)"         "GOR(L)"           
#>  [43] "LX(R)"             "AL-DM1(R)"         "AL-VC3l(R)"       
#>  [46] "AL-VP1d(R)"        "GA(R)"             "VES(R)"           
#>  [49] "a'2(R)"            "MB(+ACA)(R)"       "FBl2"             
#>  [52] "PB(R7)"            "FB"                "NO(R)"            
#>  [55] "CRE(L)"            "AL-VC5(R)"         "SMP(R)"           
#>  [58] "AL-DP1m(L)"        "INP"               "AL-DM2(R)"        
#>  [61] "gL(R)"             "AL-DC3(R)"         "EBr2r4"           
#>  [64] "LX(L)"             "SAD(-AMMC)"        "SMP(L)"           
#>  [67] "SIP(R)"            "AL-DA3(R)"         "AL-VL1(R)"        
#>  [70] "g2(R)"             "bL(R)"             "CRE(R)"           
#>  [73] "AL-DL4(L)"         "CX"                "PB(L3)"           
#>  [76] "PED(R)"            "AL-DL2d(R)"        "PB(R8)"           
#>  [79] "FBl5"              "AL-VA1v(R)"        "mALT(L)"          
#>  [82] "gL(L)"             "a3(R)"             "AL-VA6(R)"        
#>  [85] "hemibrain"         "PB(L7)"            "GF(R)"            
#>  [88] "CRE(-RUB)(L)"      "g5(R)"             "PB(L4)"           
#>  [91] "AL-DA4l(R)"        "PB(L1)"            "AL-VM7v(L)"       
#>  [94] "AL-DA1(R)"         "mALT(R)"           "FBl3"             
#>  [97] "EPA(R)"            "PB"                "SNP(R)"           
#> [100] "LO(R)"             "AL-VA7l(R)"        "AL-VM2(R)"        
#> [103] "LAL(-GA)(R)"       "AL-VA6(L)"         "AL-DM3(R)"        
#> [106] "g3(R)"             "AL-DM3(L)"         "AL-DL3(R)"        
#> [109] "AL-VM3(R)"         "SCL(L)"            "bL(L)"            
#> [112] "FBl7"              "AVLP(R)"           "PENP"             
#> [115] "SPS(L)"            "AL-VA5(R)"         "AL-VP5(R)"        
#> [118] "PB(R5)"            "AB(L)"             "EBr3pw"           
#> [121] "AL-VA1d(R)"        "CAN(R)"            "AL-D(R)"          
#> [124] "EB"                "IPS(R)"            "AL-DM1(L)"        
#> [127] "AL-VA7m(R)"        "b2(R)"             "NO1(R)"           
#> [130] "AL-VP2(R)"         "AOT(R)"            "AL-DM6(R)"        
#> [133] "AL-VA2(R)"         "FB-column3"        "FBl6"             
#> [136] "b'L(L)"            "OL(R)"             "AL-VP1m(R)"       
#> [139] "b'L(R)"            "FBl4"              "FBl8"             
#> [142] "AL-VA3(R)"         "CRE(-ROB,-RUB)(R)" "AL-DA2(L)"        
#> [145] "SNP(L)"            "EBr3d"             "AL-DP1l(R)"       
#> [148] "a2(R)"             "AL-DM5(R)"         "PB(L5)"           
#> [151] "PB(R2)"            "EBr3am"            "AL-VM7d(R)"       
#> [154] "b'1(R)"            "EPA(L)"            "AL-VC3m(R)"       
#> [157] "AL-VM5d(R)"        "VLNP(R)"           "IB"               
#> [160] "GC"                "VMNP"              "RUB(L)"           
#> [163] "NO2(L)"            "AOTU(R)"           "ROB(R)"           
#> [166] "AL-D(L)"           "b1(R)"             "SPS(R)"           
#> [169] "g1(R)"             "PB(L2)"            "LOP(R)"           
#> [172] "FBl1"              "AME(R)"            "AL-DC2(R)"        
#> [175] "AL-VM4(R)"         "a'3(R)"            "AL-DM5(L)"        
#> [178] "aL(L)"             "BU(L)"             "LH(R)"            
#> [181] "AL-DL1(R)"         "AL-DL5(R)"         "AL-VP4(R)"        
#> [184] "CA(L)"             "PB(R6)"            "SAD"              
#> [187] "SLP(R)"            "AL-VC4(R)"         "PB(L9)"           
#> [190] "AL-DL4(R)"         "GOR(R)"            "NO(L)"            
#> [193] "AL-V(R)"           "POC"               "RUB(R)"           
#> [196] "ICL(R)"            "AL-VM7v(R)"        "NO"               
#> [199] "FBl9"              "AL-VM1(R)"         "BU(R)"            
#> [202] "AL-DM2(L)"         "MB(R)"             "AL-VP3(R)"        
#> [205] "AL(L)"             "AL-VA4(R)"         "AB(R)"            
#> [208] "NO2(R)"            "GNG"               "a'L(L)"           
#> [211] "AL-DA4m(R)"        "AL-DC1(R)"         "PRW"              
#> [214] "SCL(R)"            "AL-VP1l(R)"        "FLA(R)"           
#> [217] "VES(L)"            "PB(R4)"            "AL-VC2(R)"        
#> [220] "MB(L)"             "NO3(L)"            "a1(R)"            
#> [223] "ME(R)"             "AL-DC1(L)"         "AL-VC1(R)"        
#> [226] "AL-VL2a(R)"        "NO1(L)"            "EBr1"             
#> [229] "AL-DC4(R)"         "vACA(R)"          
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "SPS(L)"            "CRE(-RUB)(L)"      "FBl1"             
#>   [4] "GNG"               "AL(R)"             "GA(R)"            
#>   [7] "LOP(R)"            "EPA(R)"            "AL-DM4(L)"        
#>  [10] "MB(+ACA)(R)"       "AL-DC4(R)"         "NO2(L)"           
#>  [13] "PB(L4)"            "AL-DL1(R)"         "AL-DM1(R)"        
#>  [16] "AL-VA4(R)"         "VLNP(R)"           "BU(L)"            
#>  [19] "FBl4"              "a'L(L)"            "AL-VA2(R)"        
#>  [22] "GF(R)"             "g3(R)"             "mALT(R)"          
#>  [25] "RUB(L)"            "AL-DM4(R)"         "AL-VM4(R)"        
#>  [28] "b2(R)"             "b1(R)"             "aL(L)"            
#>  [31] "AL-VP1l(R)"        "AL-VC2(R)"         "b'2(R)"           
#>  [34] "NO"                "vACA(R)"           "CRE(-ROB,-RUB)(R)"
#>  [37] "hemibrain"         "ATL(R)"            "AL-VC3l(R)"       
#>  [40] "AL-VL2p(R)"        "CX"                "VES(R)"           
#>  [43] "AL-DL5(R)"         "AL-D(R)"           "PB(L7)"           
#>  [46] "CA(L)"             "PB(L1)"            "b'L(L)"           
#>  [49] "a'3(R)"            "AL-VP1d(R)"        "a3(R)"            
#>  [52] "AL-DA4m(R)"        "AL-VA1d(R)"        "PLP(R)"           
#>  [55] "EB"                "LH(R)"             "ICL(R)"           
#>  [58] "AL-VM7v(L)"        "EBr3pw"            "INP"              
#>  [61] "PB(L3)"            "SPS(R)"            "BU(R)"            
#>  [64] "AL-VA6(L)"         "g5(R)"             "PB(R4)"           
#>  [67] "AL-DC2(R)"         "a2(R)"             "GC"               
#>  [70] "EBr5"              "a'2(R)"            "FBl8"             
#>  [73] "bL(R)"             "PENP"              "LX(R)"            
#>  [76] "AL-DC2(L)"         "PB(R1)"            "FB"               
#>  [79] "NO3(L)"            "AL-VL1(R)"         "FBl5"             
#>  [82] "AME(R)"            "AL-DA2(L)"         "AL-VA5(R)"        
#>  [85] "PB(L8)"            "PB(R9)"            "AB(R)"            
#>  [88] "SIP(L)"            "AL-DL2d(R)"        "a1(R)"            
#>  [91] "SNP(R)"            "LX(L)"             "AL-VC1(R)"        
#>  [94] "AL-VM5v(R)"        "NO3(R)"            "AL-DL4(L)"        
#>  [97] "AL-DP1m(L)"        "AL-VA7m(R)"        "AL-DM2(L)"        
#> [100] "AL-VM7d(L)"        "AL-DC3(R)"         "AL-DP1m(R)"       
#> [103] "SIP(R)"            "ICL(L)"            "FBl9"             
#> [106] "LAL(R)"            "aL(R)"             "NO1(R)"           
#> [109] "AL-VM5d(R)"        "AL-VM7v(R)"        "AL-DL4(R)"        
#> [112] "SCL(L)"            "PB(L9)"            "LO(R)"            
#> [115] "NO(L)"             "AL-DA4l(R)"        "RUB(R)"           
#> [118] "b'1(R)"            "ATL(L)"            "EBr6"             
#> [121] "OL(R)"             "AL-VC3m(R)"        "AL-VM7d(R)"       
#> [124] "EBr3am"            "AL-DM1(L)"         "PB(R2)"           
#> [127] "AL-V(R)"           "AL-VP4(R)"         "AL-VA3(R)"        
#> [130] "AL-VC4(R)"         "PB(R6)"            "FBl2"             
#> [133] "GOR(R)"            "AL-DL3(R)"         "PB(L5)"           
#> [136] "AL-DM5(R)"         "AL-VC5(R)"         "PED(R)"           
#> [139] "b'L(R)"            "g4(R)"             "ROB(R)"           
#> [142] "GOR(L)"            "NO1(L)"            "EPA(L)"           
#> [145] "PRW"               "mALT(L)"           "LAL(-GA)(R)"      
#> [148] "AL-DM6(L)"         "AL-DC1(R)"         "AL-VA1v(R)"       
#> [151] "dACA(R)"           "VES(L)"            "AL-DA3(R)"        
#> [154] "AL-VP5(R)"         "SAD"               "gL(L)"            
#> [157] "AL(L)"             "a'1(R)"            "AL-D(L)"          
#> [160] "FB-column3"        "PB(R5)"            "VMNP"             
#> [163] "AL-VL2a(R)"        "AL-DL2v(R)"        "AL-VP2(R)"        
#> [166] "AOT(R)"            "AB(L)"             "MB(R)"            
#> [169] "CRE(R)"            "AL-DM5(L)"         "PVLP(R)"          
#> [172] "FBl3"              "AL-DL5(L)"         "AL-DM3(R)"        
#> [175] "AL-VM2(R)"         "AL-VP1m(R)"        "AOTU(R)"          
#> [178] "lACA(R)"           "g2(R)"             "ME(R)"            
#> [181] "PB(L2)"            "AMMC"              "EBr1"             
#> [184] "AL-DA3(L)"         "FBl7"              "AL-DP1l(R)"       
#> [187] "PB(R8)"            "PB"                "AL-VA6(R)"        
#> [190] "AL-VM1(R)"         "IPS(R)"            "SLP(R)"           
#> [193] "a'L(R)"            "AL-DM3(L)"         "SNP(L)"           
#> [196] "PB(R3)"            "NO(R)"             "IB"               
#> [199] "gL(R)"             "AL-VA7l(R)"        "AL-VM3(R)"        
#> [202] "SMP(R)"            "AL-DA4m(L)"        "NO2(R)"           
#> [205] "AL-DA1(R)"         "AL-DC1(L)"         "bL(L)"            
#> [208] "SAD(-AMMC)"        "FBl6"              "LAL(L)"           
#> [211] "MB(L)"             "EBr2r4"            "AL-VP3(R)"        
#> [214] "POC"               "CRE(L)"            "AVLP(R)"          
#> [217] "SMP(L)"            "WED(R)"            "AL-DM2(R)"        
#> [220] "EBr3d"             "PB(L6)"            "AL-DA2(R)"        
#> [223] "AL-DM6(R)"         "PB(R7)"            "g1(R)"            
#> [226] "SCL(R)"            "CA(R)"             "FLA(R)"           
#> [229] "AL-DC4(L)"         "CAN(R)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method  path                                  
#> [1,] "POST"  "/logout"                             
#> [2,] "PATCH" "/api/*"                              
#> [3,] "GET"   "/api/npexplorer/neuronmetavals"      
#> [4,] "GET"   "/api/v:ver/npexplorer/neuronmetavals"
#> [5,] "GET"   "/api/v:ver/npexplorer/distribution"  
#> [6,] "GET"   "/api/dbmeta/version"                 
#>      name                                                                                          
#> [1,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                      
#> [2,] "github.com/labstack/echo.glob..func1"                                                        
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMetaVals-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getVersion-fm"               
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }