Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "SMP(L)"            "SIP(L)"            "FBl7"             
#>   [4] "CRE(-ROB,-RUB)(R)" "FBl4"              "mALT(L)"          
#>   [7] "b'L(R)"            "AL(L)"             "NO"               
#>  [10] "FBl9"              "SLP(R)"            "SAD"              
#>  [13] "MB(L)"             "EBr1"              "a'2(R)"           
#>  [16] "NO2(L)"            "b'2(R)"            "b2(R)"            
#>  [19] "dACA(R)"           "EBr3d"             "NO(L)"            
#>  [22] "BU(L)"             "CA(R)"             "PB(R3)"           
#>  [25] "PB(L1)"            "ICL(L)"            "SNP(L)"           
#>  [28] "LH(R)"             "PB(R4)"            "OL(R)"            
#>  [31] "a2(R)"             "GOR(L)"            "SNP(R)"           
#>  [34] "g5(R)"             "CAN(R)"            "SIP(R)"           
#>  [37] "FBl2"              "PB(L2)"            "PENP"             
#>  [40] "ME(R)"             "PB(R1)"            "AVLP(R)"          
#>  [43] "AB(R)"             "PB(L5)"            "PB(L8)"           
#>  [46] "a'1(R)"            "AME(R)"            "aL(L)"            
#>  [49] "a3(R)"             "IB"                "aL(R)"            
#>  [52] "a'L(R)"            "FBl8"              "bL(R)"            
#>  [55] "MB(R)"             "ROB(R)"            "ICL(R)"           
#>  [58] "b'L(L)"            "a1(R)"             "NO2(R)"           
#>  [61] "PB(L4)"            "AOTU(R)"           "PED(R)"           
#>  [64] "SPS(L)"            "LX(R)"             "VES(R)"           
#>  [67] "PB(L3)"            "VMNP"              "gL(L)"            
#>  [70] "PB(R8)"            "PB(R2)"            "VLNP(R)"          
#>  [73] "CX"                "bL(L)"             "VES(L)"           
#>  [76] "PB(R5)"            "WED(R)"            "NO1(L)"           
#>  [79] "FLA(R)"            "PB(L9)"            "CRE(L)"           
#>  [82] "GA(R)"             "GC"                "EB"               
#>  [85] "ATL(L)"            "lACA(R)"           "g2(R)"            
#>  [88] "gL(R)"             "MB(+ACA)(R)"       "FBl5"             
#>  [91] "hemibrain"         "INP"               "LOP(R)"           
#>  [94] "AOT(R)"            "SMP(R)"            "mALT(R)"          
#>  [97] "g4(R)"             "LO(R)"             "IPS(R)"           
#> [100] "LAL(R)"            "POC"               "b'1(R)"           
#> [103] "FB"                "EBr5"              "PB"               
#> [106] "PLP(R)"            "PVLP(R)"           "g1(R)"            
#> [109] "ATL(R)"            "b1(R)"             "vACA(R)"          
#> [112] "GOR(R)"            "NO(R)"             "EBr3pw"           
#> [115] "PB(L7)"            "SAD(-AMMC)"        "FBl3"             
#> [118] "PRW"               "EBr6"              "AB(L)"            
#> [121] "EBr3am"            "AMMC"              "LAL(L)"           
#> [124] "GF(R)"             "LAL(-GA)(R)"       "PB(R9)"           
#> [127] "SCL(L)"            "NO1(R)"            "a'L(L)"           
#> [130] "AL(R)"             "CA(L)"             "PB(R6)"           
#> [133] "a'3(R)"            "LX(L)"             "EPA(L)"           
#> [136] "g3(R)"             "PB(R7)"            "BU(R)"            
#> [139] "PB(L6)"            "NO3(L)"            "RUB(R)"           
#> [142] "SPS(R)"            "CRE(R)"            "EBr2r4"           
#> [145] "EPA(R)"            "FBl1"              "NO3(R)"           
#> [148] "SCL(R)"            "FBl6"              "RUB(L)"           
#> [151] "GNG"              
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                                
#> [1,] "OPTIONS" "/api/*"                            
#> [2,] "POST"    "/api/v:ver/npexplorer/findneurons" 
#> [3,] "POST"    "/api/v:ver/npexplorer/completeness"
#> [4,] "CONNECT" "/api/*"                            
#> [5,] "GET"     "/api/v:ver/cached/roiconnectivity" 
#> [6,] "GET"     "/api/npexplorer/neuronmeta"        
#>      name                                                                                      
#> [1,] "github.com/labstack/echo.glob..func1"                                                    
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                      
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                      
#> [4,] "github.com/labstack/echo.glob..func1"                                                    
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                      
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMeta-fm"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }