Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "SCL(R)"            "AL-DA3(L)"         "FBl2"             
#>   [4] "RUB(R)"            "IB"                "a2(R)"            
#>   [7] "AL-DC2(L)"         "AL-VC5(R)"         "AL-VM2(R)"        
#>  [10] "ROB(R)"            "PB(R4)"            "AL-D(L)"          
#>  [13] "GNG"               "a3(R)"             "AL-VP3(R)"        
#>  [16] "FBl3"              "IPS(R)"            "AL-VL2p(R)"       
#>  [19] "AL-VM7v(R)"        "AL-VP4(R)"         "mALT(L)"          
#>  [22] "b'L(R)"            "AL-DL2d(R)"        "AL-VP2(R)"        
#>  [25] "CRE(-RUB)(L)"      "AB(R)"             "CAN(R)"           
#>  [28] "AL-DC3(R)"         "SMP(R)"            "CRE(R)"           
#>  [31] "FLA(R)"            "LX(R)"             "IB112"            
#>  [34] "AL-VA7l(R)"        "PB(R7)"            "AL-DM1(L)"        
#>  [37] "AL-DA1(R)"         "AL-DC4(R)"         "AL-VA6(R)"        
#>  [40] "AL-VM4(R)"         "AL-VP1d(R)"        "AL-DA2(L)"        
#>  [43] "AVLP(R)"           "AL(R)"             "PB(L4)"           
#>  [46] "EBr6"              "LOP(R)"            "AMMC"             
#>  [49] "b2(R)"             "dACA(R)"           "AL-DM2(L)"        
#>  [52] "AL-DM3(L)"         "NO2(R)"            "AL-VC1(R)"        
#>  [55] "PB(L6)"            "AL-DM6(R)"         "AL-DL5(L)"        
#>  [58] "NO3(L)"            "LH(R)"             "BU(L)"            
#>  [61] "AL-DC1(L)"         "AL-DC4(L)"         "PB(R3)"           
#>  [64] "AL-DP1l(R)"        "aL(L)"             "VES(L)"           
#>  [67] "AL-VA1d(R)"        "CRE(L)"            "PED(R)"           
#>  [70] "PB(R8)"            "g4(R)"             "SIP(R)"           
#>  [73] "AL-DM3(R)"         "AL-VM5v(R)"        "b'1(R)"           
#>  [76] "AL-DL4(L)"         "AL-VC4(R)"         "ICL(R)"           
#>  [79] "AL-DA2(R)"         "AL-D(R)"           "SCL(L)"           
#>  [82] "AL(L)"             "AL-DM5(L)"         "AL-VA6(L)"        
#>  [85] "RUB(L)"            "AL-VL1(R)"         "AL-VM3(R)"        
#>  [88] "NO3(R)"            "AL-DA4m(L)"        "AL-DL2v(R)"       
#>  [91] "POC"               "g3(R)"             "b'L(L)"           
#>  [94] "VLNP(R)"           "SIP(L)"            "ME(R)"            
#>  [97] "EBr2r4"            "AL-DM4(R)"         "AL-VA5(R)"        
#> [100] "PB(R6)"            "LAL(-GA)(R)"       "VMNP"             
#> [103] "AL-DM6(L)"         "FBl1"              "NO(L)"            
#> [106] "AL-VL2a(R)"        "PB(L5)"            "SPS(R)"           
#> [109] "NO1(R)"            "PB"                "AL-VA4(R)"        
#> [112] "PB(R5)"            "aL(R)"             "EBr1"             
#> [115] "EPA(R)"            "g5(R)"             "AL-DM1(R)"        
#> [118] "AL-VA7m(R)"        "PVLP(R)"           "AL-VC3l(R)"       
#> [121] "MB(L)"             "AL-VP1m(R)"        "b'2(R)"           
#> [124] "PB(L3)"            "g1(R)"             "EBr3d"            
#> [127] "BU(R)"             "NO(R)"             "hemibrain"        
#> [130] "AL-DP1m(L)"        "PB(L9)"            "SNP(L)"           
#> [133] "AME(R)"            "AOTU(R)"           "SPS(L)"           
#> [136] "LAL(R)"            "EBr3am"            "SMP(L)"           
#> [139] "a'2(R)"            "ICL(L)"            "PB(L7)"           
#> [142] "LO(R)"             "FBl7"              "CA(R)"            
#> [145] "AL-VM7d(R)"        "GC"                "VES(R)"           
#> [148] "AL-VM7d(L)"        "SAD(-AMMC)"        "AL-VA3(R)"        
#> [151] "lACA(R)"           "AL-DL3(R)"         "AOT(R)"           
#> [154] "AL-DL1(R)"         "a'3(R)"            "mALT(R)"          
#> [157] "FBl8"              "AL-VP1l(R)"        "AB(L)"            
#> [160] "PB(L1)"            "MB(R)"             "LAL(L)"           
#> [163] "EBr5"              "NO"                "NO1(L)"           
#> [166] "EB"                "AL-DC1(R)"         "AL-DL5(R)"        
#> [169] "PRW"               "GOR(L)"            "FB"               
#> [172] "AL-VM7v(L)"        "FBl5"              "PENP"             
#> [175] "g2(R)"             "b1(R)"             "INP"              
#> [178] "AL-DA4m(R)"        "AL-VA2(R)"         "AL-VM1(R)"        
#> [181] "a'1(R)"            "FBl6"              "FBl9"             
#> [184] "AL-V(R)"           "AL-DA3(R)"         "AL-DL4(R)"        
#> [187] "AL-VC2(R)"         "a'L(R)"            "PB(L2)"           
#> [190] "ATL(R)"            "LX(L)"             "SLP(R)"           
#> [193] "AL-DM5(R)"         "WED(R)"            "SAD"              
#> [196] "PLP(R)"            "AL-DM4(L)"         "PB(R1)"           
#> [199] "bL(L)"             "FB-column3"        "a'L(L)"           
#> [202] "OL(R)"             "AL-DC2(R)"         "AL-VA1v(R)"       
#> [205] "AL-VM5d(R)"        "AL-VP5(R)"         "vACA(R)"          
#> [208] "a1(R)"             "GA(R)"             "SNP(R)"           
#> [211] "PB(R9)"            "FBl4"              "EPA(L)"           
#> [214] "AL-DM2(R)"         "CRE(-ROB,-RUB)(R)" "ATL(L)"           
#> [217] "NO2(L)"            "bL(R)"             "CA(L)"            
#> [220] "gL(R)"             "GOR(R)"            "PB(L8)"           
#> [223] "EBr3pw"            "MB(+ACA)(R)"       "AL-DP1m(R)"       
#> [226] "AL-VC3m(R)"        "CX"                "GF(R)"            
#> [229] "AL-DA4l(R)"        "gL(L)"             "PB(R2)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "g1(R)"             "PB(R5)"            "SIP(R)"           
#>   [4] "CA(L)"             "ATL(R)"            "CAN(R)"           
#>   [7] "aL(L)"             "vACA(R)"           "BU(L)"            
#>  [10] "CRE(L)"            "bL(L)"             "CRE(R)"           
#>  [13] "SCL(R)"            "EBr6"              "ME(R)"            
#>  [16] "LX(L)"             "aL(R)"             "NO3(R)"           
#>  [19] "EBr1"              "PENP"              "BU(R)"            
#>  [22] "CRE(-ROB,-RUB)(R)" "a'3(R)"            "MB(+ACA)(R)"      
#>  [25] "AVLP(R)"           "EPA(L)"            "dACA(R)"          
#>  [28] "FBl5"              "bL(R)"             "MB(R)"            
#>  [31] "RUB(R)"            "g3(R)"             "WED(R)"           
#>  [34] "FB"                "PB(R3)"            "PB(R4)"           
#>  [37] "POC"               "SMP(L)"            "EBr5"             
#>  [40] "NO1(R)"            "PB(L3)"            "mALT(L)"          
#>  [43] "GC"                "g2(R)"             "AOT(R)"           
#>  [46] "b2(R)"             "SNP(R)"            "a'1(R)"           
#>  [49] "AB(R)"             "b'L(L)"            "LAL(L)"           
#>  [52] "GA(R)"             "EB"                "SLP(R)"           
#>  [55] "SMP(R)"            "VMNP"              "NO3(L)"           
#>  [58] "GNG"               "NO"                "ROB(R)"           
#>  [61] "PB(R1)"            "FBl2"              "IB"               
#>  [64] "b'2(R)"            "a2(R)"             "NO1(L)"           
#>  [67] "b'1(R)"            "VES(R)"            "LH(R)"            
#>  [70] "VLNP(R)"           "IPS(R)"            "PB(L1)"           
#>  [73] "a'2(R)"            "g5(R)"             "gL(R)"            
#>  [76] "PB(R2)"            "GOR(R)"            "GOR(L)"           
#>  [79] "AL(L)"             "PB(L6)"            "b1(R)"            
#>  [82] "g4(R)"             "FBl3"              "ICL(R)"           
#>  [85] "EBr3d"             "EBr2r4"            "PB(R9)"           
#>  [88] "PVLP(R)"           "a3(R)"             "lACA(R)"          
#>  [91] "PB(R7)"            "PB"                "PB(L2)"           
#>  [94] "ATL(L)"            "AMMC"              "CX"               
#>  [97] "mALT(R)"           "FBl9"              "CA(R)"            
#> [100] "FBl6"              "RUB(L)"            "AB(L)"            
#> [103] "MB(L)"             "SPS(L)"            "gL(L)"            
#> [106] "LOP(R)"            "AOTU(R)"           "EPA(R)"           
#> [109] "SAD(-AMMC)"        "LAL(-GA)(R)"       "b'L(R)"           
#> [112] "PB(R8)"            "EBr3pw"            "a'L(L)"           
#> [115] "PB(R6)"            "PRW"               "SPS(R)"           
#> [118] "VES(L)"            "a1(R)"             "NO2(L)"           
#> [121] "PED(R)"            "INP"               "OL(R)"            
#> [124] "EBr3am"            "GF(R)"             "SAD"              
#> [127] "LAL(R)"            "PB(L5)"            "PB(L7)"           
#> [130] "FBl4"              "NO2(R)"            "NO(R)"            
#> [133] "ICL(L)"            "PB(L4)"            "FBl8"             
#> [136] "SIP(L)"            "a'L(R)"            "LO(R)"            
#> [139] "FBl7"              "SCL(L)"            "PB(L8)"           
#> [142] "SNP(L)"            "FBl1"              "NO(L)"            
#> [145] "LX(R)"             "PB(L9)"            "AME(R)"           
#> [148] "AL(R)"             "hemibrain"         "FLA(R)"           
#> [151] "PLP(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DM5(R)"         "mALT(L)"           "ROB(R)"           
#>   [4] "AL-DM3(R)"         "AL-VP1m(R)"        "CX"               
#>   [7] "MB(L)"             "FBl3"              "AVLP(R)"          
#>  [10] "ATL(L)"            "LAL(-GA)(R)"       "NO1(L)"           
#>  [13] "PB(L7)"            "NO2(R)"            "AL-DC3(R)"        
#>  [16] "AL-DC2(L)"         "NO"                "PENP"             
#>  [19] "GC"                "AL-VM5d(R)"        "PB(L5)"           
#>  [22] "PRW"               "AL-DL5(R)"         "bL(R)"            
#>  [25] "GF(R)"             "FBl2"              "AL-DA4m(R)"       
#>  [28] "AL-VM3(R)"         "EBr3pw"            "AMMC"             
#>  [31] "NO2(L)"            "VMNP"              "AL-DL4(L)"        
#>  [34] "AL-VC5(R)"         "AL-VP3(R)"         "PED(R)"           
#>  [37] "SPS(L)"            "MB(R)"             "AL-DA3(R)"        
#>  [40] "lACA(R)"           "PB(R9)"            "AL-DC4(L)"        
#>  [43] "SMP(L)"            "b2(R)"             "ICL(R)"           
#>  [46] "AL-DL2v(R)"        "AL-VA1d(R)"        "NO(R)"            
#>  [49] "SNP(L)"            "SAD(-AMMC)"        "AL-DA4l(R)"       
#>  [52] "AL-VM4(R)"         "a'1(R)"            "ATL(R)"           
#>  [55] "AL-DM4(L)"         "AL-DC1(R)"         "AL-DL3(R)"        
#>  [58] "AL-DM1(R)"         "FBl4"              "GOR(R)"           
#>  [61] "AL-VA7m(R)"        "EBr1"              "AL-DM4(R)"        
#>  [64] "EBr3am"            "EBr2r4"            "AL-DA1(R)"        
#>  [67] "AL-VC2(R)"         "AL-VC3l(R)"        "AL-VL1(R)"        
#>  [70] "NO1(R)"            "EB"                "AL-DC2(R)"        
#>  [73] "mALT(R)"           "PB(L1)"            "AL-DM5(L)"        
#>  [76] "g5(R)"             "MB(+ACA)(R)"       "AL-DM6(R)"        
#>  [79] "gL(L)"             "AL-DC1(L)"         "NO3(R)"           
#>  [82] "b1(R)"             "AL-VA3(R)"         "CRE(R)"           
#>  [85] "aL(L)"             "a'L(R)"            "PB(L4)"           
#>  [88] "AL-DP1m(R)"        "CRE(-ROB,-RUB)(R)" "AL-DM1(L)"        
#>  [91] "PB(R1)"            "PB"                "AL-VA6(R)"        
#>  [94] "AL(L)"             "gL(R)"             "FBl8"             
#>  [97] "OL(R)"             "AL-VC4(R)"         "BU(L)"            
#> [100] "g4(R)"             "CRE(-RUB)(L)"      "AB(L)"            
#> [103] "PB(R7)"            "AL-VA7l(R)"        "LH(R)"            
#> [106] "AL-VA5(R)"         "AL-DM6(L)"         "FBl5"             
#> [109] "PB(L2)"            "WED(R)"            "AL-VM7d(R)"       
#> [112] "ME(R)"             "g1(R)"             "EBr6"             
#> [115] "FBl1"              "NO(L)"             "FB"               
#> [118] "LO(R)"             "b'L(L)"            "LOP(R)"           
#> [121] "GNG"               "AL-VC1(R)"         "AL-VM1(R)"        
#> [124] "AL-V(R)"           "SIP(L)"            "AL-DA3(L)"        
#> [127] "LAL(R)"            "AL-VC3m(R)"        "PB(L6)"           
#> [130] "a3(R)"             "AL-D(R)"           "VLNP(R)"          
#> [133] "RUB(R)"            "PB(R6)"            "AL-DM2(L)"        
#> [136] "VES(L)"            "AL-VP1l(R)"        "INP"              
#> [139] "AL-VP4(R)"         "SPS(R)"            "LX(R)"            
#> [142] "PB(R5)"            "g3(R)"             "AL-VA1v(R)"       
#> [145] "PB(R4)"            "AL-VA2(R)"         "AL-DA2(L)"        
#> [148] "PB(R3)"            "AB(R)"             "a1(R)"            
#> [151] "SNP(R)"            "SIP(R)"            "VES(R)"           
#> [154] "IB"                "EBr5"              "AL-VM7d(L)"       
#> [157] "AL-VA6(L)"         "AL-VP2(R)"         "SAD"              
#> [160] "AL-DP1m(L)"        "a'3(R)"            "AL-DL4(R)"        
#> [163] "POC"               "b'2(R)"            "PVLP(R)"          
#> [166] "SMP(R)"            "EPA(R)"            "AL-VA4(R)"        
#> [169] "EBr3d"             "AL-VM7v(L)"        "NO3(L)"           
#> [172] "RUB(L)"            "SLP(R)"            "b'L(R)"           
#> [175] "a'2(R)"            "bL(L)"             "ICL(L)"           
#> [178] "a'L(L)"            "AL-VM5v(R)"        "AL-DA4m(L)"       
#> [181] "CRE(L)"            "BU(R)"             "b'1(R)"           
#> [184] "GOR(L)"            "PLP(R)"            "PB(R8)"           
#> [187] "FBl7"              "AME(R)"            "AL-VM2(R)"        
#> [190] "SCL(R)"            "PB(L8)"            "AL-VP1d(R)"       
#> [193] "FLA(R)"            "CAN(R)"            "CA(L)"            
#> [196] "PB(R2)"            "AL-DA2(R)"         "SCL(L)"           
#> [199] "LAL(L)"            "CA(R)"             "AL-VL2p(R)"       
#> [202] "AL-VP5(R)"         "AL-DL5(L)"         "EPA(L)"           
#> [205] "AOT(R)"            "hemibrain"         "AL-DL1(R)"        
#> [208] "AL-DM3(L)"         "aL(R)"             "dACA(R)"          
#> [211] "FBl6"              "FBl9"              "AL-DC4(R)"        
#> [214] "AL-DL2d(R)"        "AL-VM7v(R)"        "AOTU(R)"          
#> [217] "PB(L9)"            "a2(R)"             "AL-VL2a(R)"       
#> [220] "IPS(R)"            "LX(L)"             "AL-DM2(R)"        
#> [223] "AL-DP1l(R)"        "g2(R)"             "AL-D(L)"          
#> [226] "FB-column3"        "AL(R)"             "vACA(R)"          
#> [229] "PB(L3)"            "GA(R)"            
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "bL(L)"             "POC"               "a'2(R)"           
#>   [4] "AL-DM4(R)"         "AL-DM2(L)"         "CX"               
#>   [7] "AL-VA7l(R)"        "SAD(-AMMC)"        "AL-VP4(R)"        
#>  [10] "LAL(L)"            "PB"                "AL-VM7d(R)"       
#>  [13] "PB(L5)"            "BU(L)"             "AVLP(R)"          
#>  [16] "LAL(R)"            "AL-VC5(R)"         "SAD"              
#>  [19] "b1(R)"             "AL-VM5d(R)"        "AL-VP1m(R)"       
#>  [22] "LO(R)"             "LAL(-GA)(R)"       "AL-DM5(L)"        
#>  [25] "VES(L)"            "PB(R6)"            "SLP(R)"           
#>  [28] "GC"                "SCL(R)"            "AL-VA6(R)"        
#>  [31] "PB(R5)"            "b'1(R)"            "SCL(L)"           
#>  [34] "a'3(R)"            "CRE(-ROB,-RUB)(R)" "mALT(R)"          
#>  [37] "AL-DM4(L)"         "vACA(R)"           "b'L(R)"           
#>  [40] "AL-VA5(R)"         "AL-VA1d(R)"        "g1(R)"            
#>  [43] "AL-DA3(L)"         "GA(R)"             "AL-DL5(L)"        
#>  [46] "OL(R)"             "AL-VC3l(R)"        "AL-DM3(L)"        
#>  [49] "PB(R2)"            "AL-DC3(R)"         "AL-VC3m(R)"       
#>  [52] "IB"                "mALT(L)"           "AL-DC1(L)"        
#>  [55] "lACA(R)"           "AL-DL5(R)"         "AL-VC1(R)"        
#>  [58] "AL-VM1(R)"         "g3(R)"             "b2(R)"            
#>  [61] "PB(L9)"            "ICL(R)"            "WED(R)"           
#>  [64] "gL(L)"             "AL-D(L)"           "SIP(L)"           
#>  [67] "EBr5"              "GNG"               "hemibrain"        
#>  [70] "CRE(L)"            "EBr3pw"            "FBl3"             
#>  [73] "FBl4"              "CAN(R)"            "AL-VA7m(R)"       
#>  [76] "AMMC"              "SPS(L)"            "NO(L)"            
#>  [79] "AL-DA4l(R)"        "AL-DP1l(R)"        "VLNP(R)"          
#>  [82] "EBr2r4"            "SNP(L)"            "MB(+ACA)(R)"      
#>  [85] "AL-DC4(R)"         "dACA(R)"           "EBr3d"            
#>  [88] "EB"                "CRE(-RUB)(L)"      "AL-VL2p(R)"       
#>  [91] "AL-VM4(R)"         "CA(L)"             "GF(R)"            
#>  [94] "AL-DC2(R)"         "NO2(R)"            "AL-DM1(L)"        
#>  [97] "AL-DA3(R)"         "AL-DL4(R)"         "AL(L)"            
#> [100] "GOR(R)"            "AL-DA2(R)"         "AL-DL2v(R)"       
#> [103] "AL-VA2(R)"         "AL-VC4(R)"         "PENP"             
#> [106] "FB-column3"        "FB"                "NO1(L)"           
#> [109] "AL-VM3(R)"         "AOTU(R)"           "NO3(L)"           
#> [112] "IPS(R)"            "AL-DM3(R)"         "RUB(R)"           
#> [115] "SNP(R)"            "LX(R)"             "SIP(R)"           
#> [118] "CA(R)"             "SMP(R)"            "a2(R)"            
#> [121] "EPA(L)"            "PB(L2)"            "AME(R)"           
#> [124] "AL-VL2a(R)"        "PB(L1)"            "AL-DC2(L)"        
#> [127] "AL-VM7d(L)"        "INP"               "AL-VA1v(R)"       
#> [130] "AL-VM5v(R)"        "GOR(L)"            "AL-DP1m(L)"       
#> [133] "FBl8"              "AL-VM2(R)"         "BU(R)"            
#> [136] "FBl2"              "g2(R)"             "AL-VM7v(L)"       
#> [139] "a'1(R)"            "PB(L4)"            "AL-DL1(R)"        
#> [142] "AL-VP3(R)"         "AL-VP1l(R)"        "PLP(R)"           
#> [145] "LX(L)"             "EBr3am"            "LH(R)"            
#> [148] "AL-DM2(R)"         "AL-DC4(L)"         "AL-DM6(R)"        
#> [151] "NO2(L)"            "aL(R)"             "CRE(R)"           
#> [154] "EBr1"              "EPA(R)"            "PB(L6)"           
#> [157] "AL-DM1(R)"         "NO3(R)"            "PED(R)"           
#> [160] "PB(R8)"            "gL(R)"             "aL(L)"            
#> [163] "AL-VL1(R)"         "AL-VM7v(R)"        "PB(R7)"           
#> [166] "AL-VA4(R)"         "AL-VP1d(R)"        "PB(R3)"           
#> [169] "a'L(L)"            "AL-DL2d(R)"        "LOP(R)"           
#> [172] "PB(L8)"            "PB(R1)"            "MB(R)"            
#> [175] "AL-VC2(R)"         "MB(L)"             "NO(R)"            
#> [178] "AL-DA4m(L)"        "FBl9"              "AL-V(R)"          
#> [181] "SPS(R)"            "g5(R)"             "AL-DL3(R)"        
#> [184] "ROB(R)"            "VMNP"              "AL-DL4(L)"        
#> [187] "PB(L7)"            "b'L(L)"            "PB(L3)"           
#> [190] "FBl1"              "RUB(L)"            "PB(R9)"           
#> [193] "AL-DA2(L)"         "ICL(L)"            "SMP(L)"           
#> [196] "AL-DM5(R)"         "g4(R)"             "FBl6"             
#> [199] "AB(R)"             "AL-DP1m(R)"        "AL-D(R)"          
#> [202] "b'2(R)"            "AL-DA1(R)"         "FBl7"             
#> [205] "ATL(R)"            "AL-VP5(R)"         "AB(L)"            
#> [208] "VES(R)"            "bL(R)"             "PB(R4)"           
#> [211] "NO1(R)"            "ME(R)"             "AL(R)"            
#> [214] "EBr6"              "AL-VA3(R)"         "AL-VP2(R)"        
#> [217] "a1(R)"             "AL-DM6(L)"         "PRW"              
#> [220] "FBl5"              "a3(R)"             "AL-DA4m(R)"       
#> [223] "NO"                "AOT(R)"            "AL-VA6(L)"        
#> [226] "ATL(L)"            "PVLP(R)"           "a'L(R)"           
#> [229] "AL-DC1(R)"         "FLA(R)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method    path                                  
#> [1,] "POST"    "/api/custom/custom"                  
#> [2,] "POST"    "/api/v:ver/npexplorer/neuronmeta"    
#> [3,] "POST"    "/api/npexplorer/simpleconnections"   
#> [4,] "GET"     "/api/npexplorer/roisinneuron"        
#> [5,] "POST"    "/api/skeletons/skeleton/:dataset/:id"
#> [6,] "CONNECT" "/api/*"                              
#>      name                                                                                             
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"                  
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"     
#> [5,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                         
#> [6,] "github.com/labstack/echo.glob..func1"                                                           
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }