Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-VP1m(R)"        "bL(L)"             "INP"              
#>   [4] "AL-VC1(R)"         "lACA(R)"           "a'1(R)"           
#>   [7] "NO3(L)"            "CRE(L)"            "AL-VA5(R)"        
#>  [10] "AL-VA6(R)"         "b1(R)"             "AL-VP1l(R)"       
#>  [13] "b'L(R)"            "a'L(R)"            "AL-DC4(L)"        
#>  [16] "AL-VM7v(R)"        "AL-DM3(L)"         "bL(R)"            
#>  [19] "AL-DA4m(R)"        "AL-DP1m(L)"        "MB(+ACA)(R)"      
#>  [22] "AL-VP3(R)"         "VES(L)"            "AL-VM2(R)"        
#>  [25] "PB(L3)"            "SNP(L)"            "ATL(R)"           
#>  [28] "PB"                "AL-DM2(L)"         "ROB(R)"           
#>  [31] "PB(L9)"            "PB(R3)"            "SPS(R)"           
#>  [34] "WED(R)"            "PB(L1)"            "LX(L)"            
#>  [37] "AL-VL1(R)"         "CA(L)"             "mALT(R)"          
#>  [40] "AL-VP2(R)"         "PB(R6)"            "FBl2"             
#>  [43] "AL-DM6(R)"         "b'2(R)"            "AOTU(R)"          
#>  [46] "EBr3am"            "a'3(R)"            "GC"               
#>  [49] "FBl8"              "PB(R5)"            "AL-DM5(R)"        
#>  [52] "AL-VP1d(R)"        "PRW"               "CAN(R)"           
#>  [55] "b'L(L)"            "AL-VA7l(R)"        "AL-VM4(R)"        
#>  [58] "SCL(L)"            "AL-VM7d(L)"        "AL-DM4(R)"        
#>  [61] "AL-VA1d(R)"        "GNG"               "AL-DA4l(R)"       
#>  [64] "AL-DM3(R)"         "VLNP(R)"           "FLA(R)"           
#>  [67] "AL-DL4(L)"         "AL-DM1(L)"         "b'1(R)"           
#>  [70] "NO"                "FBl6"              "FBl9"             
#>  [73] "AL-DL4(R)"         "EBr6"              "PB(R8)"           
#>  [76] "AME(R)"            "dACA(R)"           "NO2(R)"           
#>  [79] "SIP(R)"            "PED(R)"            "EBr5"             
#>  [82] "FBl4"              "AL-DC1(R)"         "CX"               
#>  [85] "VMNP"              "ME(R)"             "AL-DC1(L)"        
#>  [88] "AB(L)"             "LX(R)"             "AL-DA3(L)"        
#>  [91] "AL-D(R)"           "AL-DL2v(R)"        "AL(L)"            
#>  [94] "ICL(L)"            "CA(R)"             "FBl5"             
#>  [97] "AL-VL2p(R)"        "AL-VM5v(R)"        "AL-D(L)"          
#> [100] "AL-VM7v(L)"        "IB112"             "AL-DA1(R)"        
#> [103] "PB(L5)"            "CRE(-ROB,-RUB)(R)" "SMP(L)"           
#> [106] "AL-DL5(L)"         "NO(R)"             "BU(L)"            
#> [109] "NO(L)"             "AL-VA2(R)"         "EBr1"             
#> [112] "SLP(R)"            "SIP(L)"            "FBl7"             
#> [115] "g1(R)"             "PB(L4)"            "AL-DL2d(R)"       
#> [118] "PB(R9)"            "FB-column3"        "AL-VM3(R)"        
#> [121] "AL-VP5(R)"         "g2(R)"             "EBr3d"            
#> [124] "AL-DM4(L)"         "PB(R7)"            "CRE(R)"           
#> [127] "PB(L8)"            "AL-VL2a(R)"        "hemibrain"        
#> [130] "NO1(L)"            "PB(L7)"            "MB(R)"            
#> [133] "SPS(L)"            "AL-DM5(L)"         "AL-VC2(R)"        
#> [136] "AL-DL3(R)"         "gL(L)"             "AL-VA6(L)"        
#> [139] "AL-DP1m(R)"        "LAL(-GA)(R)"       "PLP(R)"           
#> [142] "EB"                "PB(L6)"            "AL-VM7d(R)"       
#> [145] "ATL(L)"            "FBl1"              "PB(R4)"           
#> [148] "a'L(L)"            "AL-VC3l(R)"        "g3(R)"            
#> [151] "AL-DC2(L)"         "CRE(-RUB)(L)"      "LOP(R)"           
#> [154] "EPA(L)"            "AL-DL1(R)"         "AL-DL5(R)"        
#> [157] "g5(R)"             "IPS(R)"            "AL-VA7m(R)"       
#> [160] "AOT(R)"            "PENP"              "a2(R)"            
#> [163] "SAD"               "SCL(R)"            "BU(R)"            
#> [166] "VES(R)"            "PB(R2)"            "AL-VM1(R)"        
#> [169] "AL-VM5d(R)"        "gL(R)"             "AL-DM1(R)"        
#> [172] "AL-VA3(R)"         "vACA(R)"           "aL(R)"            
#> [175] "LH(R)"             "AVLP(R)"           "AL-VC5(R)"        
#> [178] "AL-DM6(L)"         "SAD(-AMMC)"        "AL-VC4(R)"        
#> [181] "NO3(R)"            "NO2(L)"            "AL-DC2(R)"        
#> [184] "AL-DM2(R)"         "LAL(L)"            "SNP(R)"           
#> [187] "AB(R)"             "RUB(L)"            "NO1(R)"           
#> [190] "LO(R)"             "PVLP(R)"           "AL-VA4(R)"        
#> [193] "b2(R)"             "POC"               "mALT(L)"          
#> [196] "a'2(R)"            "ICL(R)"            "AL-DA3(R)"        
#> [199] "IB"                "AL-DC3(R)"         "AL-DC4(R)"        
#> [202] "AL-VP4(R)"         "AL(R)"             "GF(R)"            
#> [205] "GA(R)"             "GOR(R)"            "GOR(L)"           
#> [208] "LAL(R)"            "AL-VA1v(R)"        "AL-V(R)"          
#> [211] "MB(L)"             "PB(R1)"            "AL-DP1l(R)"       
#> [214] "PB(L2)"            "a1(R)"             "FBl3"             
#> [217] "AL-DA2(R)"         "RUB(R)"            "AMMC"             
#> [220] "FB"                "g4(R)"             "SMP(R)"           
#> [223] "EBr3pw"            "a3(R)"             "OL(R)"            
#> [226] "EBr2r4"            "AL-DA2(L)"         "AL-DA4m(L)"       
#> [229] "EPA(R)"            "aL(L)"             "AL-VC3m(R)"       
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "a'L(L)"            "AL(R)"             "PB(L3)"           
#>   [4] "PB(L1)"            "PB(R8)"            "PB(L7)"           
#>   [7] "bL(R)"             "PB(L4)"            "POC"              
#>  [10] "a1(R)"             "PRW"               "PED(R)"           
#>  [13] "LH(R)"             "LX(R)"             "gL(L)"            
#>  [16] "NO(R)"             "CRE(L)"            "FBl6"             
#>  [19] "a3(R)"             "VMNP"              "SLP(R)"           
#>  [22] "vACA(R)"           "ATL(L)"            "PENP"             
#>  [25] "SMP(L)"            "SCL(R)"            "a'L(R)"           
#>  [28] "ICL(L)"            "SNP(L)"            "WED(R)"           
#>  [31] "PB(R3)"            "AOT(R)"            "NO1(L)"           
#>  [34] "dACA(R)"           "FBl8"              "SMP(R)"           
#>  [37] "GA(R)"             "PLP(R)"            "bL(L)"            
#>  [40] "MB(+ACA)(R)"       "FBl2"              "PB(L5)"           
#>  [43] "mALT(R)"           "EBr5"              "EBr3d"            
#>  [46] "PVLP(R)"           "lACA(R)"           "GOR(L)"           
#>  [49] "LAL(-GA)(R)"       "ME(R)"             "MB(R)"            
#>  [52] "SAD"               "SPS(R)"            "EPA(L)"           
#>  [55] "NO3(R)"            "EBr6"              "SPS(L)"           
#>  [58] "EBr2r4"            "BU(R)"             "PB(R2)"           
#>  [61] "b'1(R)"            "CX"                "PB(R7)"           
#>  [64] "g1(R)"             "ROB(R)"            "EBr3pw"           
#>  [67] "FBl5"              "PB(R4)"            "PB(R9)"           
#>  [70] "gL(R)"             "CAN(R)"            "aL(L)"            
#>  [73] "b'L(L)"            "a'3(R)"            "a2(R)"            
#>  [76] "hemibrain"         "b'L(R)"            "FBl1"             
#>  [79] "GOR(R)"            "RUB(R)"            "g3(R)"            
#>  [82] "PB(L9)"            "CA(L)"             "b'2(R)"           
#>  [85] "NO1(R)"            "EB"                "CRE(-ROB,-RUB)(R)"
#>  [88] "CRE(R)"            "g4(R)"             "PB(L6)"           
#>  [91] "SAD(-AMMC)"        "b1(R)"             "NO2(L)"           
#>  [94] "AMMC"              "SIP(L)"            "SNP(R)"           
#>  [97] "a'1(R)"            "VES(R)"            "FBl7"             
#> [100] "FBl9"              "MB(L)"             "GNG"              
#> [103] "LAL(R)"            "PB(L2)"            "GC"               
#> [106] "INP"               "AME(R)"            "OL(R)"            
#> [109] "VES(L)"            "EPA(R)"            "AOTU(R)"          
#> [112] "PB(R1)"            "NO3(L)"            "CA(R)"            
#> [115] "NO(L)"             "PB"                "AL(L)"            
#> [118] "LO(R)"             "AB(L)"             "LOP(R)"           
#> [121] "SIP(R)"            "FBl3"              "VLNP(R)"          
#> [124] "mALT(L)"           "LAL(L)"            "GF(R)"            
#> [127] "aL(R)"             "AB(R)"             "SCL(L)"           
#> [130] "b2(R)"             "FB"                "NO"               
#> [133] "ATL(R)"            "IPS(R)"            "PB(R6)"           
#> [136] "ICL(R)"            "g2(R)"             "g5(R)"            
#> [139] "AVLP(R)"           "EBr3am"            "EBr1"             
#> [142] "FLA(R)"            "NO2(R)"            "RUB(L)"           
#> [145] "PB(R5)"            "IB"                "BU(L)"            
#> [148] "PB(L8)"            "a'2(R)"            "FBl4"             
#> [151] "LX(L)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "PB(R7)"            "AL-D(L)"           "bL(L)"            
#>   [4] "b1(R)"             "AL-VM7d(R)"        "ICL(L)"           
#>   [7] "FBl1"              "GNG"               "AL-VA3(R)"        
#>  [10] "BU(L)"             "AL-DM6(L)"         "AL-DA2(R)"        
#>  [13] "MB(L)"             "NO1(R)"            "GOR(R)"           
#>  [16] "AL-VC5(R)"         "AL-V(R)"           "AL-DP1m(L)"       
#>  [19] "AL-D(R)"           "AL-VM1(R)"         "vACA(R)"          
#>  [22] "mALT(R)"           "AL-DM3(L)"         "LH(R)"            
#>  [25] "AL-DP1l(R)"        "PB(L2)"            "ATL(L)"           
#>  [28] "g2(R)"             "gL(L)"             "AL-DC4(L)"        
#>  [31] "AL-DM5(L)"         "LO(R)"             "PB(L4)"           
#>  [34] "AL-DA1(R)"         "EBr6"              "PB"               
#>  [37] "ME(R)"             "AL-DL2d(R)"        "PB(R6)"           
#>  [40] "SMP(L)"            "PB(R8)"            "g4(R)"            
#>  [43] "ATL(R)"            "IB"                "a'L(R)"           
#>  [46] "AME(R)"            "AL-VP1d(R)"        "VLNP(R)"          
#>  [49] "PB(R5)"            "lACA(R)"           "FB"               
#>  [52] "AL-DC1(R)"         "SMP(R)"            "PB(L1)"           
#>  [55] "a'L(L)"            "AOT(R)"            "b2(R)"            
#>  [58] "SCL(R)"            "AL-VM4(R)"         "AL-VP1m(R)"       
#>  [61] "LAL(L)"            "EPA(R)"            "NO3(L)"           
#>  [64] "PB(R2)"            "AL-VC3m(R)"        "SNP(R)"           
#>  [67] "AL-DM2(L)"         "AL-VM5d(R)"        "AL-VL2a(R)"       
#>  [70] "NO2(L)"            "LAL(-GA)(R)"       "EBr3d"            
#>  [73] "a'1(R)"            "NO(L)"             "AL-VA1d(R)"       
#>  [76] "FBl5"              "AL-VA7m(R)"        "AL-VC4(R)"        
#>  [79] "AL-DC1(L)"         "FBl6"              "AL-VL1(R)"        
#>  [82] "LX(R)"             "AL-VM7v(L)"        "AL-DL2v(R)"       
#>  [85] "AL-DM5(R)"         "AL-VA1v(R)"        "SAD"              
#>  [88] "NO"                "AL-DM4(R)"         "LOP(R)"           
#>  [91] "g1(R)"             "PB(R9)"            "AB(R)"            
#>  [94] "SLP(R)"            "AL-VM5v(R)"        "RUB(L)"           
#>  [97] "VES(L)"            "WED(R)"            "AL-DA2(L)"        
#> [100] "SIP(R)"            "AL-VC1(R)"         "hemibrain"        
#> [103] "aL(R)"             "IPS(R)"            "AL-VA6(R)"        
#> [106] "AL-DP1m(R)"        "RUB(R)"            "SCL(L)"           
#> [109] "EBr3am"            "EBr2r4"            "FB-column3"       
#> [112] "AB(L)"             "AL-DA3(L)"         "AOTU(R)"          
#> [115] "SPS(L)"            "CA(R)"             "CRE(R)"           
#> [118] "PB(R3)"            "AL-VP2(R)"         "POC"              
#> [121] "NO1(L)"            "AL-DM4(L)"         "AL-DL4(R)"        
#> [124] "AL-DM2(R)"         "AL-VM3(R)"         "AL-DL1(R)"        
#> [127] "AL-VM7v(R)"        "PENP"              "SIP(L)"           
#> [130] "dACA(R)"           "AL-VA2(R)"         "AL(L)"            
#> [133] "AL-DL5(L)"         "NO2(R)"            "AL-VP4(R)"        
#> [136] "FLA(R)"            "AL-DC2(L)"         "AL-VA6(L)"        
#> [139] "MB(+ACA)(R)"       "VMNP"              "EBr3pw"           
#> [142] "GF(R)"             "g3(R)"             "GC"               
#> [145] "CRE(-RUB)(L)"      "AL-DA4m(R)"        "AL-VC3l(R)"       
#> [148] "AL-VP1l(R)"        "AMMC"              "AL(R)"            
#> [151] "AL-VA7l(R)"        "PB(L3)"            "NO(R)"            
#> [154] "SNP(L)"            "FBl4"              "CAN(R)"           
#> [157] "PED(R)"            "a'2(R)"            "PB(L7)"           
#> [160] "g5(R)"             "b'L(L)"            "b'2(R)"           
#> [163] "AL-DL4(L)"         "FBl3"              "b'L(R)"           
#> [166] "PLP(R)"            "AL-VM7d(L)"        "AL-DL3(R)"        
#> [169] "a1(R)"             "EBr1"              "PB(L6)"           
#> [172] "a3(R)"             "PB(L5)"            "PRW"              
#> [175] "AL-DM3(R)"         "SPS(R)"            "FBl2"             
#> [178] "AL-VA5(R)"         "MB(R)"             "mALT(L)"          
#> [181] "PB(R1)"            "EPA(L)"            "bL(R)"            
#> [184] "AL-DC4(R)"         "GA(R)"             "PB(L9)"           
#> [187] "AL-VM2(R)"         "a'3(R)"            "CRE(-ROB,-RUB)(R)"
#> [190] "AL-DA3(R)"         "AL-DL5(R)"         "AL-VA4(R)"        
#> [193] "NO3(R)"            "PB(L8)"            "FBl7"             
#> [196] "FBl8"              "LAL(R)"            "AL-DA4m(L)"       
#> [199] "ICL(R)"            "ROB(R)"            "AL-DM1(R)"        
#> [202] "b'1(R)"            "EB"                "AL-DC2(R)"        
#> [205] "AL-VP5(R)"         "CA(L)"             "CRE(L)"           
#> [208] "PVLP(R)"           "FBl9"              "INP"              
#> [211] "PB(R4)"            "AL-DM1(L)"         "AL-DA4l(R)"       
#> [214] "AL-VL2p(R)"        "EBr5"              "AVLP(R)"          
#> [217] "AL-VP3(R)"         "CX"                "BU(R)"            
#> [220] "VES(R)"            "AL-VC2(R)"         "a2(R)"            
#> [223] "SAD(-AMMC)"        "OL(R)"             "AL-DC3(R)"        
#> [226] "gL(R)"             "GOR(L)"            "LX(L)"            
#> [229] "aL(L)"             "AL-DM6(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "PVLP(R)"           "AL-VC3m(R)"        "AL-D(R)"          
#>   [4] "AL-VP4(R)"         "SPS(R)"            "dACA(R)"          
#>   [7] "AL-DA3(L)"         "PB(L7)"            "FBl7"             
#>  [10] "AL-VM5d(R)"        "VES(R)"            "AL-VP1m(R)"       
#>  [13] "VES(L)"            "AL-VP5(R)"         "g2(R)"            
#>  [16] "EB"                "NO"                "FBl1"             
#>  [19] "AL-DM1(R)"         "CA(L)"             "EBr3pw"           
#>  [22] "AL(R)"             "hemibrain"         "SAD"              
#>  [25] "SPS(L)"            "AL-VM7v(L)"        "PB(R6)"           
#>  [28] "VMNP"              "LH(R)"             "AL(L)"            
#>  [31] "FBl6"              "AL-VP1l(R)"        "AL-DP1m(R)"       
#>  [34] "GF(R)"             "ME(R)"             "CAN(R)"           
#>  [37] "AL-VC4(R)"         "SMP(R)"            "AL-DA4l(R)"       
#>  [40] "PB(R7)"            "FB"                "vACA(R)"          
#>  [43] "a1(R)"             "b'2(R)"            "SNP(L)"           
#>  [46] "AL-VC5(R)"         "AL-VA2(R)"         "AL-VA7m(R)"       
#>  [49] "CRE(-ROB,-RUB)(R)" "GA(R)"             "LAL(R)"           
#>  [52] "AL-VL1(R)"         "b'1(R)"            "lACA(R)"          
#>  [55] "AL-DA2(L)"         "SIP(R)"            "FBl2"             
#>  [58] "AL-VA3(R)"         "AL-VP3(R)"         "AL-DC3(R)"        
#>  [61] "AL-VC1(R)"         "AL-VM7v(R)"        "CRE(R)"           
#>  [64] "NO(R)"             "AL-DL3(R)"         "EPA(L)"           
#>  [67] "AL-DM4(R)"         "AL-DP1l(R)"        "AL-VC3l(R)"       
#>  [70] "SLP(R)"            "AL-DL4(R)"         "LAL(L)"           
#>  [73] "AL-DL5(L)"         "AL-VA6(L)"         "gL(R)"            
#>  [76] "LX(L)"             "PB(L4)"            "PENP"             
#>  [79] "MB(R)"             "AL-VA7l(R)"        "SCL(L)"           
#>  [82] "mALT(L)"           "EBr1"              "AL-DM6(L)"        
#>  [85] "PB(R4)"            "AL-VM4(R)"         "g1(R)"            
#>  [88] "NO3(L)"            "AL-VC2(R)"         "VLNP(R)"          
#>  [91] "FBl9"              "CA(R)"             "AL-VA5(R)"        
#>  [94] "POC"               "PB(L1)"            "PLP(R)"           
#>  [97] "AL-DA2(R)"         "IB"                "AMMC"             
#> [100] "CX"                "EBr6"              "SNP(R)"           
#> [103] "GOR(R)"            "OL(R)"             "mALT(R)"          
#> [106] "a'2(R)"            "ICL(R)"            "PB(R5)"           
#> [109] "EBr3d"             "gL(L)"             "AL-DC1(L)"        
#> [112] "AVLP(R)"           "AL-VA1v(R)"        "AOT(R)"           
#> [115] "RUB(R)"            "BU(R)"             "AL-D(L)"          
#> [118] "AL-DM4(L)"         "PB(L9)"            "PB(L5)"           
#> [121] "PRW"               "BU(L)"             "PB(R9)"           
#> [124] "b'L(L)"            "AL-DC2(R)"         "AL-VM5v(R)"       
#> [127] "aL(R)"             "CRE(-RUB)(L)"      "g5(R)"            
#> [130] "ATL(R)"            "aL(L)"             "AL-VP2(R)"        
#> [133] "AL-DM2(L)"         "LOP(R)"            "g3(R)"            
#> [136] "b'L(R)"            "GC"                "PB"               
#> [139] "RUB(L)"            "AL-VL2a(R)"        "a'3(R)"           
#> [142] "AL-DL4(L)"         "PB(R8)"            "AL-DM5(L)"        
#> [145] "AL-DM5(R)"         "AL-VA1d(R)"        "AL-VM3(R)"        
#> [148] "LAL(-GA)(R)"       "AL-DP1m(L)"        "AB(R)"            
#> [151] "a3(R)"             "AL-DL2v(R)"        "EBr5"             
#> [154] "AL-DC4(L)"         "FBl5"              "PB(R2)"           
#> [157] "SCL(R)"            "AL-VM7d(R)"        "FLA(R)"           
#> [160] "AL-DA4m(L)"        "bL(L)"             "AL-DM1(L)"        
#> [163] "FB-column3"        "INP"               "AME(R)"           
#> [166] "NO(L)"             "AL-VP1d(R)"        "GOR(L)"           
#> [169] "EBr2r4"            "ICL(L)"            "FBl4"             
#> [172] "ATL(L)"            "AL-DM3(R)"         "AL-DC2(L)"        
#> [175] "ROB(R)"            "FBl3"              "b1(R)"            
#> [178] "AL-DL1(R)"         "AL-DL2d(R)"        "MB(L)"            
#> [181] "g4(R)"             "SAD(-AMMC)"        "PB(L8)"           
#> [184] "AL-VM7d(L)"        "AL-DC1(R)"         "SMP(L)"           
#> [187] "WED(R)"            "b2(R)"             "a'1(R)"           
#> [190] "LO(R)"             "NO1(R)"            "AL-DM6(R)"        
#> [193] "PB(R3)"            "IPS(R)"            "NO2(L)"           
#> [196] "PB(L6)"            "GNG"               "PB(L2)"           
#> [199] "MB(+ACA)(R)"       "AB(L)"             "PB(L3)"           
#> [202] "NO3(R)"            "a'L(R)"            "a'L(L)"           
#> [205] "AL-DC4(R)"         "AL-DL5(R)"         "SIP(L)"           
#> [208] "bL(R)"             "AL-VL2p(R)"        "AL-V(R)"          
#> [211] "EBr3am"            "AOTU(R)"           "a2(R)"            
#> [214] "LX(R)"             "FBl8"              "NO2(R)"           
#> [217] "PED(R)"            "PB(R1)"            "CRE(L)"           
#> [220] "AL-DA1(R)"         "AL-DA3(R)"         "AL-VA4(R)"        
#> [223] "NO1(L)"            "EPA(R)"            "AL-DM2(R)"        
#> [226] "AL-VM1(R)"         "AL-VM2(R)"         "AL-DM3(L)"        
#> [229] "AL-DA4m(R)"        "AL-VA6(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method    path                                
#> [1,] "GET"     "/api/available"                    
#> [2,] "HEAD"    "/api"                              
#> [3,] "CONNECT" "/api/*"                            
#> [4,] "POST"    "/api/v:ver/npexplorer/rankedtable" 
#> [5,] "POST"    "/api/v:ver/npexplorer/completeness"
#> [6,] "GET"     "/api/v:ver/dbmeta/datasets"        
#>      name                                                                
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.SetupRoutes.func1" 
#> [2,] "github.com/labstack/echo/v4.glob..func1"                           
#> [3,] "github.com/labstack/echo/v4.glob..func1"                           
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }