Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "BU(L)"             "AL-DM1(R)"         "AL-VA3(R)"        
#>   [4] "EPA(R)"            "AL-VA1v(R)"        "SCL(R)"           
#>   [7] "BU(R)"             "AL-DL4(L)"         "AL-VM2(R)"        
#>  [10] "AL-DC2(L)"         "AL-DM6(L)"         "gL(R)"            
#>  [13] "FBl3"              "FB"                "LX(R)"            
#>  [16] "vACA(R)"           "LAL(L)"            "AL-DA4m(R)"       
#>  [19] "SCL(L)"            "PB(R8)"            "EBr3am"           
#>  [22] "PRW"               "AL-D(R)"           "AL-VC5(R)"        
#>  [25] "AL-VM4(R)"         "AL-VC4(R)"         "AL-VM5v(R)"       
#>  [28] "GF(R)"             "SIP(R)"            "SAD"              
#>  [31] "PB(L9)"            "AME(R)"            "b'2(R)"           
#>  [34] "MB(L)"             "EBr1"              "PB(L2)"           
#>  [37] "VES(L)"            "AL-VM1(R)"         "AL-V(R)"          
#>  [40] "AMMC"              "AL-DL5(R)"         "AL-VC3l(R)"       
#>  [43] "PB(L7)"            "FBl4"              "AL-DC3(R)"        
#>  [46] "AL-DL2d(R)"        "PB(L4)"            "AL-DA1(R)"        
#>  [49] "AL-VA7m(R)"        "PB(R5)"            "aL(R)"            
#>  [52] "a'L(R)"            "INP"               "SLP(R)"           
#>  [55] "PB(R6)"            "FLA(R)"            "ICL(R)"           
#>  [58] "CX"                "a'2(R)"            "CA(R)"            
#>  [61] "AL-VM7d(R)"        "AL-VM3(R)"         "AL-DM5(L)"        
#>  [64] "AL-DL4(R)"         "AL-DM5(R)"         "AL-VA1d(R)"       
#>  [67] "ATL(R)"            "b'L(L)"            "AL-DC2(R)"        
#>  [70] "g3(R)"             "NO"                "CRE(-ROB,-RUB)(R)"
#>  [73] "NO1(R)"            "g5(R)"             "SNP(L)"           
#>  [76] "RUB(R)"            "GOR(L)"            "AL-DM2(L)"        
#>  [79] "EPA(L)"            "GOR(R)"            "CA(L)"            
#>  [82] "AL-DL5(L)"         "AL-VM7v(L)"        "AL-VC1(R)"        
#>  [85] "AL-VC3m(R)"        "AL-DA3(R)"         "AL-VC2(R)"        
#>  [88] "dACA(R)"           "PB"                "NO3(L)"           
#>  [91] "GNG"               "EBr3pw"            "MB(+ACA)(R)"      
#>  [94] "AL-VM5d(R)"        "PENP"              "b2(R)"            
#>  [97] "PB(L8)"            "SPS(L)"            "AL-DL1(R)"        
#> [100] "a2(R)"             "PB(R7)"            "EBr3d"            
#> [103] "AL(R)"             "AL-DP1l(R)"        "EBr6"             
#> [106] "ROB(R)"            "CAN(R)"            "AOT(R)"           
#> [109] "SNP(R)"            "LH(R)"             "AL-VA2(R)"        
#> [112] "g1(R)"             "EBr2r4"            "AL-DC4(L)"        
#> [115] "hemibrain"         "FBl5"              "AL-VA5(R)"        
#> [118] "a1(R)"             "AOTU(R)"           "AL-DM1(L)"        
#> [121] "AL-DC1(R)"         "AL-DM3(R)"         "mALT(R)"          
#> [124] "AL-VM7d(L)"        "PB(R9)"            "AL-VP1l(R)"       
#> [127] "PB(L3)"            "AL-DA3(L)"         "PB(R1)"           
#> [130] "CRE(L)"            "VMNP"              "AL-DA2(R)"        
#> [133] "AL-DL3(R)"         "CRE(R)"            "bL(R)"            
#> [136] "MB(R)"             "SPS(R)"            "IB112"            
#> [139] "NO(L)"             "AL-DL2v(R)"        "b1(R)"            
#> [142] "GA(R)"             "g4(R)"             "AL-DC1(L)"        
#> [145] "ICL(L)"            "AL-VP1d(R)"        "NO3(R)"           
#> [148] "AL-DA4m(L)"        "AL-VA4(R)"         "AL-VP3(R)"        
#> [151] "ATL(L)"            "SMP(L)"            "FBl7"             
#> [154] "FBl8"              "a'3(R)"            "PB(R2)"           
#> [157] "LX(L)"             "LOP(R)"            "SIP(L)"           
#> [160] "WED(R)"            "IB"                "PB(L5)"           
#> [163] "FBl2"              "SMP(R)"            "mALT(L)"          
#> [166] "PB(L1)"            "AL-VP1m(R)"        "VLNP(R)"          
#> [169] "POC"               "NO2(L)"            "AL-VA6(L)"        
#> [172] "AB(R)"             "AL-VA7l(R)"        "AL-VL1(R)"        
#> [175] "PB(L6)"            "AL-DP1m(R)"        "AL-VL2p(R)"       
#> [178] "b'1(R)"            "ME(R)"             "gL(L)"            
#> [181] "PVLP(R)"           "PB(R4)"            "GC"               
#> [184] "FB-column3"        "IPS(R)"            "RUB(L)"           
#> [187] "AL-DC4(R)"         "AL-VM7v(R)"        "AL-VP2(R)"        
#> [190] "NO1(L)"            "b'L(R)"            "g2(R)"            
#> [193] "EBr5"              "NO(R)"             "LO(R)"            
#> [196] "AL-DA4l(R)"        "AB(L)"             "AL-VA6(R)"        
#> [199] "lACA(R)"           "CRE(-RUB)(L)"      "NO2(R)"           
#> [202] "AL-VP4(R)"         "FBl9"              "aL(L)"            
#> [205] "OL(R)"             "PLP(R)"            "AL-DM3(L)"        
#> [208] "a'1(R)"            "FBl6"              "AL-DA2(L)"        
#> [211] "a3(R)"             "AL-DM6(R)"         "PED(R)"           
#> [214] "LAL(-GA)(R)"       "SAD(-AMMC)"        "FBl1"             
#> [217] "AL-DM2(R)"         "AL-D(L)"           "AL-DP1m(L)"       
#> [220] "VES(R)"            "AL-DM4(R)"         "AL(L)"            
#> [223] "bL(L)"             "a'L(L)"            "EB"               
#> [226] "PB(R3)"            "LAL(R)"            "AL-VL2a(R)"       
#> [229] "AL-DM4(L)"         "AVLP(R)"           "AL-VP5(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "CRE(R)"            "PB(R8)"            "ICL(L)"           
#>   [4] "NO(L)"             "a1(R)"             "GA(R)"            
#>   [7] "FBl1"              "SMP(L)"            "NO3(R)"           
#>  [10] "SPS(L)"            "PB"                "gL(R)"            
#>  [13] "AVLP(R)"           "a3(R)"             "MB(L)"            
#>  [16] "PB(L9)"            "PB(R6)"            "PRW"              
#>  [19] "FLA(R)"            "EPA(R)"            "EPA(L)"           
#>  [22] "LOP(R)"            "g4(R)"             "EBr3d"            
#>  [25] "NO3(L)"            "GNG"               "GOR(R)"           
#>  [28] "NO2(L)"            "g3(R)"             "SPS(R)"           
#>  [31] "LAL(-GA)(R)"       "PB(R9)"            "CA(R)"            
#>  [34] "CAN(R)"            "RUB(L)"            "SCL(R)"           
#>  [37] "bL(L)"             "FBl5"              "IPS(R)"           
#>  [40] "PB(R5)"            "b'1(R)"            "AMMC"             
#>  [43] "NO1(R)"            "CRE(L)"            "PB(L7)"           
#>  [46] "VES(R)"            "PB(R2)"            "LAL(R)"           
#>  [49] "EBr6"              "MB(+ACA)(R)"       "AL(L)"            
#>  [52] "g5(R)"             "MB(R)"             "PENP"             
#>  [55] "PB(R7)"            "PB(L1)"            "NO"               
#>  [58] "SLP(R)"            "SIP(L)"            "lACA(R)"          
#>  [61] "dACA(R)"           "LX(R)"             "SIP(R)"           
#>  [64] "NO2(R)"            "GF(R)"             "PLP(R)"           
#>  [67] "gL(L)"             "a'1(R)"            "FBl4"             
#>  [70] "ATL(R)"            "ICL(R)"            "GOR(L)"           
#>  [73] "aL(R)"             "NO(R)"             "FBl6"             
#>  [76] "PB(L2)"            "AB(L)"             "RUB(R)"           
#>  [79] "hemibrain"         "PB(L6)"            "SAD(-AMMC)"       
#>  [82] "POC"               "g1(R)"             "EBr2r4"           
#>  [85] "EB"                "PB(R1)"            "AB(R)"            
#>  [88] "VES(L)"            "VMNP"              "PVLP(R)"          
#>  [91] "AL(R)"             "EBr3pw"            "AOT(R)"           
#>  [94] "a'3(R)"            "mALT(L)"           "a2(R)"            
#>  [97] "LH(R)"             "PB(R3)"            "INP"              
#> [100] "VLNP(R)"           "CRE(-ROB,-RUB)(R)" "PB(L4)"           
#> [103] "mALT(R)"           "b2(R)"             "EBr5"             
#> [106] "a'2(R)"            "PB(R4)"            "bL(R)"            
#> [109] "b'L(L)"            "IB"                "PB(L5)"           
#> [112] "CA(L)"             "SAD"               "PB(L8)"           
#> [115] "BU(R)"             "AME(R)"            "SMP(R)"           
#> [118] "b'2(R)"            "PED(R)"            "EBr1"             
#> [121] "b1(R)"             "AOTU(R)"           "GC"               
#> [124] "WED(R)"            "ROB(R)"            "FBl7"             
#> [127] "LX(L)"             "LAL(L)"            "NO1(L)"           
#> [130] "ME(R)"             "a'L(R)"            "OL(R)"            
#> [133] "EBr3am"            "PB(L3)"            "FB"               
#> [136] "b'L(R)"            "FBl2"              "ATL(L)"           
#> [139] "BU(L)"             "FBl3"              "FBl8"             
#> [142] "FBl9"              "a'L(L)"            "vACA(R)"          
#> [145] "g2(R)"             "SNP(R)"            "SNP(L)"           
#> [148] "LO(R)"             "aL(L)"             "SCL(L)"           
#> [151] "CX"               
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "PED(R)"            "PB(L4)"            "AL-VA7l(R)"       
#>   [4] "AL-DM2(L)"         "AL-DM4(R)"         "AL-VC4(R)"        
#>   [7] "NO1(R)"            "AL-DC3(R)"         "AL-DM1(R)"        
#>  [10] "AL-VM7v(R)"        "PB(L3)"            "aL(R)"            
#>  [13] "PB(L6)"            "b2(R)"             "AB(L)"            
#>  [16] "IB"                "g1(R)"             "POC"              
#>  [19] "dACA(R)"           "AL-D(L)"           "LAL(R)"           
#>  [22] "AL-VC1(R)"         "vACA(R)"           "SMP(L)"           
#>  [25] "gL(R)"             "AL-VP1m(R)"        "AL-VM5d(R)"       
#>  [28] "NO3(R)"            "PLP(R)"            "PB(R1)"           
#>  [31] "AME(R)"            "OL(R)"             "AL-DP1m(R)"       
#>  [34] "PENP"              "EBr3pw"            "GOR(R)"           
#>  [37] "NO2(L)"            "SMP(R)"            "AL-DC2(R)"        
#>  [40] "AL-VL2p(R)"        "AL-VP1l(R)"        "PB(R5)"           
#>  [43] "LOP(R)"            "GOR(L)"            "BU(L)"            
#>  [46] "EBr3d"             "LH(R)"             "EPA(L)"           
#>  [49] "AL-VA4(R)"         "PB(R8)"            "RUB(L)"           
#>  [52] "AL-DA1(R)"         "AL-DM5(L)"         "PB(R7)"           
#>  [55] "VES(R)"            "a2(R)"             "PB(L7)"           
#>  [58] "CAN(R)"            "a'3(R)"            "CA(L)"            
#>  [61] "GA(R)"             "SAD"               "EBr6"             
#>  [64] "AL(L)"             "AL-DC4(L)"         "AL-DM3(R)"        
#>  [67] "GC"                "FB"                "g5(R)"            
#>  [70] "AL-VC3m(R)"        "AL-VM3(R)"         "ATL(L)"           
#>  [73] "a'2(R)"            "IPS(R)"            "AL-DC1(R)"        
#>  [76] "AL-DL2v(R)"        "AL-VP4(R)"         "AMMC"             
#>  [79] "PB(R2)"            "ATL(R)"            "AL-DL5(R)"        
#>  [82] "EBr1"              "AL-VL2a(R)"        "AL-VP2(R)"        
#>  [85] "VLNP(R)"           "ME(R)"             "EB"               
#>  [88] "AL-DM3(L)"         "INP"               "AL-VL1(R)"        
#>  [91] "AL-VM7d(R)"        "CRE(-ROB,-RUB)(R)" "BU(R)"            
#>  [94] "EPA(R)"            "PB(L9)"            "b1(R)"            
#>  [97] "SCL(L)"            "MB(L)"             "AL-DM6(L)"        
#> [100] "ICL(R)"            "AL-DC4(R)"         "PB(R4)"           
#> [103] "VES(L)"            "AL-VA2(R)"         "AL-VA6(R)"        
#> [106] "PRW"               "SCL(R)"            "b'1(R)"           
#> [109] "SLP(R)"            "AL-DM6(R)"         "PB(L5)"           
#> [112] "LX(R)"             "FB-column3"        "AL-VA5(R)"        
#> [115] "LAL(-GA)(R)"       "NO(L)"             "CRE(R)"           
#> [118] "AL-DL4(L)"         "GNG"               "CX"               
#> [121] "AL-DA4m(L)"        "AL-VM7v(L)"        "PB(R3)"           
#> [124] "a3(R)"             "AL-DL2d(R)"        "AL-VA7m(R)"       
#> [127] "AL-VM4(R)"         "GF(R)"             "EBr2r4"           
#> [130] "AL-DC1(L)"         "a'1(R)"            "b'L(L)"           
#> [133] "AL-VA1d(R)"        "AL-DM1(L)"         "FBl8"             
#> [136] "AL-VM1(R)"         "VMNP"              "FBl4"             
#> [139] "FBl6"              "aL(L)"             "AL-VA1v(R)"       
#> [142] "AL-VC5(R)"         "lACA(R)"           "mALT(R)"          
#> [145] "MB(R)"             "AL(R)"             "AL-DA4m(R)"       
#> [148] "AL-D(R)"           "ROB(R)"            "FBl5"             
#> [151] "AL-DL3(R)"         "g3(R)"             "PB(L8)"           
#> [154] "FBl2"              "AL-DA3(R)"         "AOT(R)"           
#> [157] "PB(L2)"            "a'L(R)"            "AL-VA6(L)"        
#> [160] "NO"                "NO2(R)"            "RUB(R)"           
#> [163] "AL-VP1d(R)"        "AL-DP1m(L)"        "ICL(L)"           
#> [166] "bL(R)"             "AL-DA2(R)"         "AL-VA3(R)"        
#> [169] "AL-VM7d(L)"        "CA(R)"             "AL-VM5v(R)"       
#> [172] "AVLP(R)"           "NO1(L)"            "FBl1"             
#> [175] "AL-VP3(R)"         "hemibrain"         "gL(L)"            
#> [178] "MB(+ACA)(R)"       "AB(R)"             "LX(L)"            
#> [181] "AL-VP5(R)"         "b'2(R)"            "LAL(L)"           
#> [184] "FLA(R)"            "LO(R)"             "FBl3"             
#> [187] "FBl9"              "SPS(R)"            "EBr5"             
#> [190] "CRE(-RUB)(L)"      "SNP(L)"            "AL-VC2(R)"        
#> [193] "AL-V(R)"           "AL-DL4(R)"         "AL-VM2(R)"        
#> [196] "b'L(R)"            "g2(R)"             "AL-DL5(L)"        
#> [199] "WED(R)"            "CRE(L)"            "bL(L)"            
#> [202] "FBl7"              "AL-DA2(L)"         "AL-DM5(R)"        
#> [205] "AL-DM4(L)"         "AL-DA4l(R)"        "AL-VC3l(R)"       
#> [208] "AOTU(R)"           "PB"                "AL-DL1(R)"        
#> [211] "mALT(L)"           "PB(L1)"            "AL-DA3(L)"        
#> [214] "SIP(R)"            "PB(R9)"            "AL-DP1l(R)"       
#> [217] "EBr3am"            "a1(R)"             "SIP(L)"           
#> [220] "SPS(L)"            "SNP(R)"            "NO(R)"            
#> [223] "SAD(-AMMC)"        "a'L(L)"            "PB(R6)"           
#> [226] "NO3(L)"            "PVLP(R)"           "g4(R)"            
#> [229] "AL-DC2(L)"         "AL-DM2(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "MB(R)"             "AL-VA5(R)"         "AL-DL5(L)"        
#>   [4] "AL-VP3(R)"         "FBl6"              "AL-DL3(R)"        
#>   [7] "b'1(R)"            "IB"                "LX(R)"            
#>  [10] "AL-VM5d(R)"        "AL-DL5(R)"         "bL(L)"            
#>  [13] "IPS(R)"            "PB(R4)"            "AL-DL1(R)"        
#>  [16] "EBr5"              "LAL(L)"            "ME(R)"            
#>  [19] "RUB(L)"            "PB(L4)"            "AL-VC5(R)"        
#>  [22] "AOT(R)"            "PB(R6)"            "CRE(-RUB)(L)"     
#>  [25] "PVLP(R)"           "AL-VP1d(R)"        "EBr3d"            
#>  [28] "AL-VL2p(R)"        "a'3(R)"            "NO2(R)"           
#>  [31] "AL-DC2(R)"         "gL(L)"             "CRE(L)"           
#>  [34] "EBr2r4"            "a'1(R)"            "FBl7"             
#>  [37] "CA(L)"             "NO3(R)"            "FBl1"             
#>  [40] "FB-column3"        "FBl9"              "AL-DM5(R)"        
#>  [43] "AL-DM6(R)"         "g4(R)"             "CA(R)"            
#>  [46] "AVLP(R)"           "b'L(L)"            "a'L(L)"           
#>  [49] "GA(R)"             "dACA(R)"           "PB"               
#>  [52] "SNP(R)"            "AL-DA2(L)"         "PB(L2)"           
#>  [55] "b'L(R)"            "AL-DP1m(L)"        "AL-VA6(L)"        
#>  [58] "AL-DC4(R)"         "AL-DM1(R)"         "PB(L3)"           
#>  [61] "AL-DM4(L)"         "PB(L8)"            "AL-VL2a(R)"       
#>  [64] "LAL(R)"            "AB(L)"             "SCL(L)"           
#>  [67] "a1(R)"             "AL-VA3(R)"         "PRW"              
#>  [70] "EBr6"              "NO2(L)"            "AOTU(R)"          
#>  [73] "SIP(R)"            "AL-DL2d(R)"        "LAL(-GA)(R)"      
#>  [76] "PB(R9)"            "ATL(R)"            "AL-VL1(R)"        
#>  [79] "AL-VM7d(R)"        "AL-VP1l(R)"        "SCL(R)"           
#>  [82] "AL-D(L)"           "NO(R)"             "a2(R)"            
#>  [85] "AL-DM3(R)"         "AL-VC1(R)"         "NO3(L)"           
#>  [88] "bL(R)"             "a3(R)"             "AL-DC3(R)"        
#>  [91] "NO1(L)"            "GC"                "g2(R)"            
#>  [94] "VMNP"              "LO(R)"             "b'2(R)"           
#>  [97] "FBl5"              "PB(R2)"            "CAN(R)"           
#> [100] "g3(R)"             "EBr1"              "BU(L)"            
#> [103] "SLP(R)"            "AL(R)"             "b2(R)"            
#> [106] "aL(L)"             "AL-DA3(R)"         "AL-DP1l(R)"       
#> [109] "AL-VM3(R)"         "SMP(R)"            "NO1(R)"           
#> [112] "SAD"               "FLA(R)"            "AL-DA1(R)"        
#> [115] "AL-VA2(R)"         "EBr3am"            "SMP(L)"           
#> [118] "AL-DM1(L)"         "b1(R)"             "AL-DC1(R)"        
#> [121] "VLNP(R)"           "CX"                "PLP(R)"           
#> [124] "AL-DM2(R)"         "VES(L)"            "RUB(R)"           
#> [127] "AMMC"              "SPS(L)"            "AL-DA4m(L)"       
#> [130] "AL-VM7v(L)"        "g5(R)"             "AME(R)"           
#> [133] "PB(R7)"            "AL-DL4(L)"         "LH(R)"            
#> [136] "AL-VA1v(R)"        "AL-VP2(R)"         "AL-DC1(L)"        
#> [139] "AL-DA4l(R)"        "MB(+ACA)(R)"       "NO"               
#> [142] "GNG"               "AL-DA2(R)"         "g1(R)"            
#> [145] "AL-VC3l(R)"        "PB(R5)"            "ATL(L)"           
#> [148] "GOR(L)"            "aL(R)"             "PB(R1)"           
#> [151] "PB(L9)"            "ICL(L)"            "LX(L)"            
#> [154] "POC"               "SIP(L)"            "MB(L)"            
#> [157] "FBl2"              "AL-D(R)"           "AL-VP4(R)"        
#> [160] "WED(R)"            "VES(R)"            "FBl8"             
#> [163] "AL-VC4(R)"         "GF(R)"             "INP"              
#> [166] "hemibrain"         "EB"                "EBr3pw"           
#> [169] "PB(R8)"            "LOP(R)"            "CRE(R)"           
#> [172] "GOR(R)"            "AL-DL2v(R)"        "AL-DA3(L)"        
#> [175] "mALT(R)"           "PB(L1)"            "gL(R)"            
#> [178] "FBl3"              "NO(L)"             "SPS(R)"           
#> [181] "AL(L)"             "AL-VM7d(L)"        "ICL(R)"           
#> [184] "AL-VA6(R)"         "FB"                "CRE(-ROB,-RUB)(R)"
#> [187] "a'2(R)"            "PED(R)"            "AL-DC4(L)"        
#> [190] "AL-DM3(L)"         "FBl4"              "lACA(R)"          
#> [193] "PB(L5)"            "AL-VA7m(R)"        "AL-DM2(L)"        
#> [196] "PB(L7)"            "EPA(L)"            "AL-V(R)"          
#> [199] "AL-VP5(R)"         "AL-DM6(L)"         "AL-DM4(R)"        
#> [202] "AL-VA1d(R)"        "AL-VM1(R)"         "AL-VM2(R)"        
#> [205] "a'L(R)"            "AL-VA7l(R)"        "vACA(R)"          
#> [208] "AL-DM5(L)"         "PB(R3)"            "AL-DL4(R)"        
#> [211] "AL-VM7v(R)"        "PENP"              "AL-DP1m(R)"       
#> [214] "AL-VC2(R)"         "AL-VM5v(R)"        "OL(R)"            
#> [217] "ROB(R)"            "PB(L6)"            "AL-VA4(R)"        
#> [220] "AL-VC3m(R)"        "AL-VP1m(R)"        "AL-DC2(L)"        
#> [223] "SAD(-AMMC)"        "BU(R)"             "SNP(L)"           
#> [226] "AB(R)"             "AL-DA4m(R)"        "AL-VM4(R)"        
#> [229] "mALT(L)"           "EPA(R)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method  path                                  
#> [1,] "GET"   "/api/serverinfo"                     
#> [2,] "GET"   "/*"                                  
#> [3,] "POST"  "/logout"                             
#> [4,] "PATCH" "/api/*"                              
#> [5,] "GET"   "/api/npexplorer/neuronmetavals"      
#> [6,] "GET"   "/api/v:ver/npexplorer/neuronmetavals"
#>      name                                                                                          
#> [1,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                      
#> [2,] "github.com/labstack/echo.common.static.func1"                                                
#> [3,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                      
#> [4,] "github.com/labstack/echo.glob..func1"                                                        
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMetaVals-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }