Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "mALT(R)"           "PB(L1)"            "NO2(R)"           
#>   [4] "hemibrain"         "NO1(L)"            "FBl6"             
#>   [7] "FBl9"              "b'1(R)"            "AOTU(R)"          
#>  [10] "ATL(R)"            "ICL(R)"            "bL(R)"            
#>  [13] "AL(R)"             "PB(L5)"            "VES(R)"           
#>  [16] "PB(L4)"            "a1(R)"             "PENP"             
#>  [19] "NO3(R)"            "ICL(L)"            "PB(L7)"           
#>  [22] "CAN(R)"            "PB(L2)"            "PED(R)"           
#>  [25] "PB(R9)"            "ROB(R)"            "b'L(L)"           
#>  [28] "IB"                "SCL(L)"            "SIP(R)"           
#>  [31] "PB(R8)"            "EBr2r4"            "NO"               
#>  [34] "PB(L9)"            "LH(R)"             "OL(R)"            
#>  [37] "PB(R6)"            "SNP(L)"            "LAL(R)"           
#>  [40] "SMP(L)"            "EBr1"              "ME(R)"            
#>  [43] "EPA(R)"            "NO3(L)"            "AVLP(R)"          
#>  [46] "aL(L)"             "GA(R)"             "b'L(R)"           
#>  [49] "FLA(R)"            "SAD(-AMMC)"        "LO(R)"            
#>  [52] "GNG"               "POC"               "g3(R)"            
#>  [55] "PB(L8)"            "AOT(R)"            "AB(L)"            
#>  [58] "a2(R)"             "g2(R)"             "EBr6"             
#>  [61] "FBl1"              "CA(L)"             "SIP(L)"           
#>  [64] "a'2(R)"            "FBl2"              "MB(R)"            
#>  [67] "VLNP(R)"           "LOP(R)"            "a'3(R)"           
#>  [70] "aL(R)"             "EBr3d"             "INP"              
#>  [73] "GOR(L)"            "b2(R)"             "FBl4"             
#>  [76] "FBl8"              "GF(R)"             "LX(R)"            
#>  [79] "AL(L)"             "PB(R2)"            "LX(L)"            
#>  [82] "PB(R4)"            "mALT(L)"           "SAD"              
#>  [85] "bL(L)"             "a3(R)"             "SLP(R)"           
#>  [88] "MB(L)"             "dACA(R)"           "FBl7"             
#>  [91] "CX"                "GC"                "gL(L)"            
#>  [94] "BU(R)"             "AB(R)"             "ATL(L)"           
#>  [97] "g1(R)"             "NO1(R)"            "EPA(L)"           
#> [100] "NO2(L)"            "CRE(R)"            "VES(L)"           
#> [103] "AME(R)"            "PB(R7)"            "IPS(R)"           
#> [106] "AMMC"              "LAL(L)"            "MB(+ACA)(R)"      
#> [109] "PB(R3)"            "PLP(R)"            "PVLP(R)"          
#> [112] "WED(R)"            "FB"                "SPS(L)"           
#> [115] "CRE(L)"            "FBl3"              "SPS(R)"           
#> [118] "g4(R)"             "a'1(R)"            "g5(R)"            
#> [121] "FBl5"              "GOR(R)"            "PB(R5)"           
#> [124] "EBr5"              "BU(L)"             "a'L(R)"           
#> [127] "lACA(R)"           "PB(L3)"            "VMNP"             
#> [130] "NO(L)"             "SMP(R)"            "RUB(R)"           
#> [133] "b'2(R)"            "EBr3pw"            "RUB(L)"           
#> [136] "vACA(R)"           "SCL(R)"            "SNP(R)"           
#> [139] "PB(R1)"            "a'L(L)"            "CRE(-ROB,-RUB)(R)"
#> [142] "PRW"               "PB(L6)"            "gL(R)"            
#> [145] "CA(R)"             "LAL(-GA)(R)"       "EB"               
#> [148] "NO(R)"             "b1(R)"             "EBr3am"           
#> [151] "PB"               
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                        
#> [1,] "GET"  "/api/v:ver/npexplorer/findneurons"         
#> [2,] "GET"  "/api/v:ver/skeletons/skeleton/:dataset/:id"
#> [3,] "POST" "/api/v:ver/raw/cypher/transaction"         
#> [4,] "GET"  "/api/v:ver/raw/keyvalue/key/:instance/:key"
#> [5,] "GET"  "/api/available"                            
#> [6,] "POST" "/api/npexplorer/rankedtable"               
#>      name                                                                                       
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.SetupRoutes.func1"                        
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getRankedTable-fm"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }