Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AVLP(R)"           "RUB(L)"            "AL-DP1l(R)"       
#>   [4] "FBl2"              "CA(R)"             "AL-DM6(R)"        
#>   [7] "AL-VM5v(R)"        "FBl4"              "AL-VA1v(R)"       
#>  [10] "AL-DA4m(R)"        "EBr6"              "AL-VM7d(L)"       
#>  [13] "a'L(L)"            "SAD"               "AL-VP2(R)"        
#>  [16] "FBl5"              "ICL(R)"            "AL-VA1d(R)"       
#>  [19] "AL-VC5(R)"         "AL-VP5(R)"         "ATL(L)"           
#>  [22] "PB(L3)"            "MB(R)"             "AL-DA4l(R)"       
#>  [25] "AL-DL2v(R)"        "AL-VA5(R)"         "AL-VM5d(R)"       
#>  [28] "IPS(R)"            "AL-DA4m(L)"        "LX(L)"            
#>  [31] "AL-DM4(R)"         "SPS(L)"            "SCL(R)"           
#>  [34] "AL-DC2(L)"         "NO3(R)"            "aL(L)"            
#>  [37] "EB"                "FBl3"              "VES(L)"           
#>  [40] "EBr3d"             "GOR(R)"            "AL-DA3(R)"        
#>  [43] "AL-DC2(R)"         "AL-VM2(R)"         "CRE(-ROB,-RUB)(R)"
#>  [46] "AL-DL5(L)"         "GNG"               "AL-DL4(R)"        
#>  [49] "AL-DM3(R)"         "AL-VM3(R)"         "RUB(R)"           
#>  [52] "PB(L6)"            "CAN(R)"            "AL-DC3(R)"        
#>  [55] "CRE(L)"            "PB(L1)"            "AB(L)"            
#>  [58] "SNP(L)"            "PB(L8)"            "PB(R9)"           
#>  [61] "PB(R7)"            "PB(L9)"            "FBl6"             
#>  [64] "AOT(R)"            "AMMC"              "VMNP"             
#>  [67] "PB(R3)"            "AL-DM5(L)"         "NO"               
#>  [70] "OL(R)"             "lACA(R)"           "BU(L)"            
#>  [73] "EPA(R)"            "AL-DA2(L)"         "AL-VA2(R)"        
#>  [76] "IB"                "LAL(L)"            "AL-DA3(L)"        
#>  [79] "AL-DL5(R)"         "SPS(R)"            "AL-DM5(R)"        
#>  [82] "AL-VP1m(R)"        "PED(R)"            "AL-VA6(L)"        
#>  [85] "AL-VL2a(R)"        "AL-VM7v(R)"        "PB(R5)"           
#>  [88] "FB"                "a'L(R)"            "a3(R)"            
#>  [91] "AL-DL2d(R)"        "b'L(R)"            "ICL(L)"           
#>  [94] "NO(L)"             "AL-DM1(R)"         "AL-VP3(R)"        
#>  [97] "SMP(R)"            "NO(R)"             "AL-D(L)"          
#> [100] "FBl9"              "AL-D(R)"           "a1(R)"            
#> [103] "GOR(L)"            "AL-DC1(L)"         "AB(R)"            
#> [106] "AL-VA6(R)"         "AL-VC2(R)"         "WED(R)"           
#> [109] "ATL(R)"            "AL-DA1(R)"         "b'2(R)"           
#> [112] "PB(R1)"            "AL-DM2(L)"         "AL-DL4(L)"        
#> [115] "AL-VL1(R)"         "vACA(R)"           "PB(R6)"           
#> [118] "AOTU(R)"           "AL-VL2p(R)"        "g3(R)"            
#> [121] "AL-VC1(R)"         "PENP"              "AL-VP4(R)"        
#> [124] "mALT(L)"           "AL-VA3(R)"         "MB(L)"            
#> [127] "a2(R)"             "AL-DM3(L)"         "PB(R2)"           
#> [130] "SLP(R)"            "AL-VA4(R)"         "GF(R)"            
#> [133] "LH(R)"             "MB(+ACA)(R)"       "LAL(-GA)(R)"      
#> [136] "EBr5"              "ROB(R)"            "g5(R)"            
#> [139] "PRW"               "PB(L5)"            "AL(R)"            
#> [142] "a'2(R)"            "PB(L7)"            "IB112"            
#> [145] "PVLP(R)"           "VLNP(R)"           "hemibrain"        
#> [148] "CRE(-RUB)(L)"      "AL-VM7d(R)"        "AL-VP1l(R)"       
#> [151] "dACA(R)"           "CRE(R)"            "AL-VM7v(L)"       
#> [154] "AL-DM6(L)"         "FLA(R)"            "SMP(L)"           
#> [157] "FB-column3"        "FBl8"              "AL-DP1m(L)"       
#> [160] "SIP(R)"            "LAL(R)"            "CX"               
#> [163] "AL(L)"             "b'1(R)"            "NO1(L)"           
#> [166] "BU(R)"             "AL-DM1(L)"         "AL-VC3l(R)"       
#> [169] "AL-V(R)"           "AL-VP1d(R)"        "NO2(L)"           
#> [172] "GC"                "ME(R)"             "EBr3pw"           
#> [175] "FBl7"              "INP"               "SIP(L)"           
#> [178] "SCL(L)"            "LX(R)"             "PB(L4)"           
#> [181] "AL-DL3(R)"         "AL-DM2(R)"         "b2(R)"            
#> [184] "a'3(R)"            "a'1(R)"            "LO(R)"            
#> [187] "AL-VA7l(R)"        "AL-VC4(R)"         "g2(R)"            
#> [190] "g1(R)"             "gL(L)"             "g4(R)"            
#> [193] "SAD(-AMMC)"        "POC"               "GA(R)"            
#> [196] "FBl1"              "bL(R)"             "bL(L)"            
#> [199] "NO2(R)"            "AL-DC1(R)"         "AL-DL1(R)"        
#> [202] "AL-DP1m(R)"        "AL-VC3m(R)"        "AL-VM1(R)"        
#> [205] "PB(L2)"            "SNP(R)"            "b1(R)"            
#> [208] "AL-DA2(R)"         "CA(L)"             "mALT(R)"          
#> [211] "AL-VM4(R)"         "EBr3am"            "NO3(L)"           
#> [214] "NO1(R)"            "PB(R8)"            "gL(R)"            
#> [217] "VES(R)"            "PB(R4)"            "LOP(R)"           
#> [220] "EBr1"              "AL-VA7m(R)"        "AL-DC4(L)"        
#> [223] "PB"                "EBr2r4"            "EPA(L)"           
#> [226] "b'L(L)"            "AL-DC4(R)"         "aL(R)"            
#> [229] "AL-DM4(L)"         "AME(R)"            "PLP(R)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "WED(R)"            "PB(L8)"            "FBl2"             
#>   [4] "AB(R)"             "a1(R)"             "POC"              
#>   [7] "PB(L2)"            "PB"                "g5(R)"            
#>  [10] "PB(R2)"            "b1(R)"             "AME(R)"           
#>  [13] "GC"                "PB(L9)"            "ICL(R)"           
#>  [16] "IPS(R)"            "LOP(R)"            "PB(L5)"           
#>  [19] "SMP(L)"            "MB(L)"             "SNP(L)"           
#>  [22] "AL(R)"             "RUB(R)"            "BU(R)"            
#>  [25] "EPA(R)"            "INP"               "CX"               
#>  [28] "PB(R7)"            "PLP(R)"            "AB(L)"            
#>  [31] "EPA(L)"            "GOR(R)"            "PENP"             
#>  [34] "LX(R)"             "ME(R)"             "MB(+ACA)(R)"      
#>  [37] "FBl6"              "PB(L7)"            "bL(R)"            
#>  [40] "NO(L)"             "SIP(L)"            "g1(R)"            
#>  [43] "CRE(R)"            "b2(R)"             "b'L(R)"           
#>  [46] "PB(R5)"            "NO3(R)"            "EBr5"             
#>  [49] "SPS(L)"            "PB(L6)"            "a3(R)"            
#>  [52] "EBr3pw"            "ICL(L)"            "PB(R3)"           
#>  [55] "CRE(-ROB,-RUB)(R)" "LX(L)"             "AOTU(R)"          
#>  [58] "SPS(R)"            "a'L(R)"            "NO"               
#>  [61] "PVLP(R)"           "a2(R)"             "hemibrain"        
#>  [64] "FLA(R)"            "a'L(L)"            "lACA(R)"          
#>  [67] "FBl5"              "NO2(R)"            "OL(R)"            
#>  [70] "MB(R)"             "FBl1"              "LH(R)"            
#>  [73] "ATL(R)"            "SLP(R)"            "PB(L3)"           
#>  [76] "VES(R)"            "VLNP(R)"           "SCL(L)"           
#>  [79] "mALT(L)"           "EBr6"              "NO3(L)"           
#>  [82] "GNG"               "ATL(L)"            "SAD"              
#>  [85] "ROB(R)"            "PB(R1)"            "FBl8"             
#>  [88] "FBl9"              "PB(R9)"            "CAN(R)"           
#>  [91] "GOR(L)"            "LAL(-GA)(R)"       "FB"               
#>  [94] "SCL(R)"            "NO1(R)"            "g4(R)"            
#>  [97] "AL(L)"             "IB"                "SNP(R)"           
#> [100] "a'1(R)"            "FBl3"              "b'1(R)"           
#> [103] "FBl4"              "EBr3am"            "CA(L)"            
#> [106] "VES(L)"            "aL(R)"             "EB"               
#> [109] "gL(R)"             "LO(R)"             "PB(R4)"           
#> [112] "bL(L)"             "AOT(R)"            "g3(R)"            
#> [115] "VMNP"              "gL(L)"             "SIP(R)"           
#> [118] "CRE(L)"            "SAD(-AMMC)"        "GF(R)"            
#> [121] "GA(R)"             "NO1(L)"            "BU(L)"            
#> [124] "PB(L1)"            "AVLP(R)"           "RUB(L)"           
#> [127] "PRW"               "NO(R)"             "vACA(R)"          
#> [130] "b'2(R)"            "g2(R)"             "EBr3d"            
#> [133] "b'L(L)"            "NO2(L)"            "PB(R6)"           
#> [136] "EBr1"              "EBr2r4"            "FBl7"             
#> [139] "PB(R8)"            "a'2(R)"            "PB(L4)"           
#> [142] "a'3(R)"            "aL(L)"             "LAL(R)"           
#> [145] "AMMC"              "SMP(R)"            "LAL(L)"           
#> [148] "PED(R)"            "mALT(R)"           "dACA(R)"          
#> [151] "CA(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "LO(R)"             "LAL(-GA)(R)"       "AL(L)"            
#>   [4] "a3(R)"             "SIP(L)"            "AL-DM5(L)"        
#>   [7] "AL-VM7d(R)"        "CA(L)"             "FLA(R)"           
#>  [10] "AVLP(R)"           "LX(L)"             "AL(R)"            
#>  [13] "SIP(R)"            "b2(R)"             "EBr6"             
#>  [16] "a'2(R)"            "AL-DC2(L)"         "AL-VL2a(R)"       
#>  [19] "PENP"              "FBl2"              "ICL(R)"           
#>  [22] "AL-DC4(L)"         "NO1(R)"            "EPA(R)"           
#>  [25] "IB"                "SAD"               "PB(L2)"           
#>  [28] "aL(L)"             "a1(R)"             "AL-VA6(R)"        
#>  [31] "IPS(R)"            "SLP(R)"            "NO3(R)"           
#>  [34] "SNP(L)"            "EB"                "NO2(R)"           
#>  [37] "GOR(R)"            "AL-VA3(R)"         "AL-VM2(R)"        
#>  [40] "hemibrain"         "MB(+ACA)(R)"       "LAL(R)"           
#>  [43] "AL-DC1(R)"         "AL-DC2(R)"         "SAD(-AMMC)"       
#>  [46] "ROB(R)"            "EBr3pw"            "AL-VC2(R)"        
#>  [49] "AL-VP3(R)"         "PB(L3)"            "AL-D(L)"          
#>  [52] "AL-DA2(R)"         "AL-DM6(R)"         "AL-VC1(R)"        
#>  [55] "AL-DA3(L)"         "SMP(L)"            "CRE(R)"           
#>  [58] "ME(R)"             "bL(R)"             "AL-DL3(R)"        
#>  [61] "AL-VC4(R)"         "PB(R6)"            "FBl6"             
#>  [64] "PED(R)"            "PB(L7)"            "AL-DL4(R)"        
#>  [67] "SPS(R)"            "AB(L)"             "FB-column3"       
#>  [70] "PB(R1)"            "AL-DM1(R)"         "AL-VL2p(R)"       
#>  [73] "AL-VP2(R)"         "b'2(R)"            "AL-VP4(R)"        
#>  [76] "NO1(L)"            "AL-DA4m(L)"        "AL-DA4l(R)"       
#>  [79] "AL-DL2v(R)"        "NO2(L)"            "lACA(R)"          
#>  [82] "a2(R)"             "GC"                "NO(L)"            
#>  [85] "AL-DM3(R)"         "AL-DP1m(R)"        "AL-VA7l(R)"       
#>  [88] "EBr3am"            "LAL(L)"            "a'3(R)"           
#>  [91] "AL-DL5(L)"         "CRE(L)"            "FBl8"             
#>  [94] "MB(R)"             "PB(R7)"            "dACA(R)"          
#>  [97] "AME(R)"            "GF(R)"             "VES(L)"           
#> [100] "AL-VC3l(R)"        "CX"                "AL-DM3(L)"        
#> [103] "b1(R)"             "CAN(R)"            "AL-DL5(R)"        
#> [106] "AL-VA1v(R)"        "AL-VM1(R)"         "a'1(R)"           
#> [109] "BU(R)"             "VES(R)"            "AMMC"             
#> [112] "g1(R)"             "CA(R)"             "AB(R)"            
#> [115] "AL-VA7m(R)"        "RUB(R)"            "g4(R)"            
#> [118] "g5(R)"             "AL-DC3(R)"         "AL-DC1(L)"        
#> [121] "NO3(L)"            "RUB(L)"            "AL-VC5(R)"        
#> [124] "b'1(R)"            "aL(R)"             "AOTU(R)"          
#> [127] "a'L(R)"            "bL(L)"             "ATL(R)"           
#> [130] "g3(R)"             "AL-D(R)"           "AL-VM4(R)"        
#> [133] "AL-DA4m(R)"        "PB(L4)"            "PB(R5)"           
#> [136] "vACA(R)"           "NO(R)"             "LX(R)"            
#> [139] "AL-DP1m(L)"        "AL-DC4(R)"         "PB"               
#> [142] "SMP(R)"            "WED(R)"            "EBr1"             
#> [145] "PB(L1)"            "AL-VP1d(R)"        "POC"              
#> [148] "SCL(L)"            "PB(L5)"            "mALT(R)"          
#> [151] "AL-DL4(L)"         "b'L(L)"            "SPS(L)"           
#> [154] "EBr5"              "a'L(L)"            "AL-DA3(R)"        
#> [157] "AL-DL2d(R)"        "PB(R8)"            "VMNP"             
#> [160] "b'L(R)"            "FB"                "AL-DM2(R)"        
#> [163] "AL-VM7v(R)"        "gL(R)"             "AL-VM5v(R)"       
#> [166] "PVLP(R)"           "AL-VP5(R)"         "PRW"              
#> [169] "LH(R)"             "ICL(L)"            "FBl5"             
#> [172] "VLNP(R)"           "SCL(R)"            "AL-V(R)"          
#> [175] "LOP(R)"            "AL-VM7d(L)"        "AL-DM4(R)"        
#> [178] "AL-VA5(R)"         "PB(R3)"            "PB(L9)"           
#> [181] "FBl3"              "mALT(L)"           "AL-DM1(L)"        
#> [184] "AL-VM7v(L)"        "AL-DA1(R)"         "AL-DA2(L)"        
#> [187] "PB(R4)"            "AL-DP1l(R)"        "ATL(L)"           
#> [190] "GA(R)"             "AL-VM3(R)"         "CRE(-RUB)(L)"     
#> [193] "EBr3d"             "SNP(R)"            "NO"               
#> [196] "PB(L6)"            "AL-VL1(R)"         "AOT(R)"           
#> [199] "PB(L8)"            "CRE(-ROB,-RUB)(R)" "BU(L)"            
#> [202] "FBl1"              "GNG"               "AL-VA2(R)"        
#> [205] "AL-VC3m(R)"        "AL-VP1m(R)"        "GOR(L)"           
#> [208] "AL-VA6(L)"         "FBl9"              "AL-DM5(R)"        
#> [211] "AL-DM6(L)"         "gL(L)"             "AL-DM2(L)"        
#> [214] "PB(R9)"            "FBl4"              "g2(R)"            
#> [217] "PLP(R)"            "EBr2r4"            "AL-DM4(L)"        
#> [220] "OL(R)"             "AL-VA4(R)"         "AL-VM5d(R)"       
#> [223] "MB(L)"             "AL-DL1(R)"         "EPA(L)"           
#> [226] "PB(R2)"            "AL-VA1d(R)"        "FBl7"             
#> [229] "AL-VP1l(R)"        "INP"              
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "PB(L8)"            "a'L(R)"            "AL-DC1(R)"        
#>   [4] "LH(R)"             "AL-DM2(R)"         "AL-DA4m(L)"       
#>   [7] "AL-VL1(R)"         "EBr1"              "ROB(R)"           
#>  [10] "FBl1"              "AL-DP1l(R)"        "PLP(R)"           
#>  [13] "MB(+ACA)(R)"       "CAN(R)"            "PB(L5)"           
#>  [16] "dACA(R)"           "PB(R1)"            "GNG"              
#>  [19] "AL-DA2(R)"         "CX"                "PB(L1)"           
#>  [22] "SIP(R)"            "FB-column3"        "AL-VM7v(R)"       
#>  [25] "lACA(R)"           "g2(R)"             "NO(L)"            
#>  [28] "AL-VC5(R)"         "FB"                "EBr2r4"           
#>  [31] "FLA(R)"            "EBr3d"             "IPS(R)"           
#>  [34] "AL-DM1(R)"         "GOR(L)"            "AL-DM5(L)"        
#>  [37] "AL-DL2v(R)"        "AL-VA7l(R)"        "AL-VP1l(R)"       
#>  [40] "EBr6"              "NO"                "FBl3"             
#>  [43] "EPA(L)"            "PB(R5)"            "GA(R)"            
#>  [46] "AL-VC2(R)"         "PB(L7)"            "AL-DP1m(R)"       
#>  [49] "EBr3am"            "PRW"               "ME(R)"            
#>  [52] "vACA(R)"           "mALT(L)"           "NO1(R)"           
#>  [55] "AL-DL4(L)"         "SNP(L)"            "bL(R)"            
#>  [58] "AL-DC4(R)"         "a'3(R)"            "LAL(L)"           
#>  [61] "FBl6"              "a'1(R)"            "AVLP(R)"          
#>  [64] "AL-VA7m(R)"        "SNP(R)"            "FBl8"             
#>  [67] "AL-DL2d(R)"        "AL-VL2p(R)"        "AL-VM1(R)"        
#>  [70] "ATL(L)"            "AL-DM4(L)"         "AL-VA2(R)"        
#>  [73] "AL-VC3m(R)"        "EBr5"              "VES(R)"           
#>  [76] "FBl5"              "AL-VC3l(R)"        "RUB(R)"           
#>  [79] "PB(R6)"            "AL(R)"             "AL-VP5(R)"        
#>  [82] "SMP(R)"            "g3(R)"             "hemibrain"        
#>  [85] "aL(R)"             "b'L(R)"            "AL-DC3(R)"        
#>  [88] "AL-VL2a(R)"        "PB(L2)"            "VMNP"             
#>  [91] "a'2(R)"            "EPA(R)"            "AL-VA6(L)"        
#>  [94] "PB(L9)"            "AL-D(L)"           "AL-DM6(R)"        
#>  [97] "AL-VM2(R)"         "AL-DM2(L)"         "LAL(R)"           
#> [100] "AL-D(R)"           "AL-VA1d(R)"        "AL-VM5d(R)"       
#> [103] "NO2(L)"            "SPS(R)"            "g1(R)"            
#> [106] "gL(L)"             "b1(R)"             "AL-VC1(R)"        
#> [109] "BU(R)"             "AL-VC4(R)"         "PENP"             
#> [112] "PB(R7)"            "CRE(L)"            "PB(L4)"           
#> [115] "VLNP(R)"           "AB(L)"             "LAL(-GA)(R)"      
#> [118] "mALT(R)"           "RUB(L)"            "AL-DM5(R)"        
#> [121] "AL-VA1v(R)"        "a1(R)"             "PB(L3)"           
#> [124] "PB"                "PB(R4)"            "b'L(L)"           
#> [127] "ATL(R)"            "MB(R)"             "AL-DM6(L)"        
#> [130] "FBl9"              "AL-DL3(R)"         "AL-VM7d(R)"       
#> [133] "SMP(L)"            "PB(L6)"            "ICL(L)"           
#> [136] "PB(R3)"            "AL-VA6(R)"         "AL-DA3(L)"        
#> [139] "NO3(L)"            "AL-DC2(R)"         "a2(R)"            
#> [142] "EBr3pw"            "AL-VP1d(R)"        "LO(R)"            
#> [145] "AME(R)"            "b2(R)"             "AL-DC1(L)"        
#> [148] "gL(R)"             "NO2(R)"            "AL-DL5(R)"        
#> [151] "AL-VM3(R)"         "AL-DC2(L)"         "CRE(-RUB)(L)"     
#> [154] "GOR(R)"            "SCL(L)"            "AOTU(R)"          
#> [157] "CRE(R)"            "ICL(R)"            "AL-VM4(R)"        
#> [160] "AL-DC4(L)"         "AL-DL1(R)"         "g4(R)"            
#> [163] "FBl2"              "FBl4"              "CA(R)"            
#> [166] "AL-VA5(R)"         "AL-DM3(L)"         "AB(R)"            
#> [169] "AL-DA3(R)"         "AL-DA4l(R)"        "POC"              
#> [172] "AL(L)"             "AL-DA2(L)"         "AL-DP1m(L)"       
#> [175] "SIP(L)"            "SPS(L)"            "PVLP(R)"          
#> [178] "VES(L)"            "AL-VA4(R)"         "b'1(R)"           
#> [181] "g5(R)"             "AL-VA3(R)"         "AL-VP3(R)"        
#> [184] "PB(R9)"            "AL-DM3(R)"         "AL-V(R)"          
#> [187] "AL-DL5(L)"         "SAD(-AMMC)"        "SAD"              
#> [190] "a'L(L)"            "OL(R)"             "NO(R)"            
#> [193] "PED(R)"            "PB(R2)"            "aL(L)"            
#> [196] "a3(R)"             "SLP(R)"            "AOT(R)"           
#> [199] "LOP(R)"            "INP"               "AL-DA4m(R)"       
#> [202] "IB"                "GF(R)"             "LX(L)"            
#> [205] "AL-DL4(R)"         "AL-VP1m(R)"        "MB(L)"            
#> [208] "EB"                "AL-DM1(L)"         "AL-VM5v(R)"       
#> [211] "AL-VP4(R)"         "SCL(R)"            "bL(L)"            
#> [214] "AL-DA1(R)"         "AL-VP2(R)"         "b'2(R)"           
#> [217] "GC"                "NO3(R)"            "NO1(L)"           
#> [220] "BU(L)"             "PB(R8)"            "FBl7"             
#> [223] "CRE(-ROB,-RUB)(R)" "CA(L)"             "WED(R)"           
#> [226] "LX(R)"             "AL-VM7d(L)"        "AL-VM7v(L)"       
#> [229] "AL-DM4(R)"         "AMMC"             
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                        
#> [1,] "GET"  "/api/npexplorer/simpleconnections"         
#> [2,] "HEAD" "/api"                                      
#> [3,] "POST" "/api/raw/cypher/cypher"                    
#> [4,] "GET"  "/api/v:ver/raw/keyvalue/key/:instance/:key"
#> [5,] "GET"  "/api/v:ver/cached/roiconnectivity"         
#> [6,] "GET"  "/api/v:ver/npexplorer/neuronmetavals"      
#>      name                                                                                             
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
#> [2,] "github.com/labstack/echo.glob..func1"                                                           
#> [3,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                         
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }