Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "GOR(L)"            "LO(R)"             "FBl6"             
#>   [4] "FBl7"              "ICL(R)"            "PB(R5)"           
#>   [7] "SMP(R)"            "RUB(R)"            "mALT(R)"          
#>  [10] "aL(R)"             "SIP(R)"            "NO3(L)"           
#>  [13] "MB(R)"             "AMMC"              "LX(R)"            
#>  [16] "INP"               "PB(R4)"            "EBr3am"           
#>  [19] "PB"                "PLP(R)"            "EBr2r4"           
#>  [22] "NO(L)"             "CRE(R)"            "AL(L)"            
#>  [25] "NO"                "PENP"              "LAL(L)"           
#>  [28] "SMP(L)"            "GA(R)"             "SIP(L)"           
#>  [31] "PB(L6)"            "PVLP(R)"           "FBl2"             
#>  [34] "FBl4"              "ATL(R)"            "PRW"              
#>  [37] "PB(R8)"            "ME(R)"             "AOT(R)"           
#>  [40] "CX"                "PB(R6)"            "AOTU(R)"          
#>  [43] "b2(R)"             "NO1(R)"            "b1(R)"            
#>  [46] "SLP(R)"            "ATL(L)"            "GF(R)"            
#>  [49] "PED(R)"            "GOR(R)"            "LAL(R)"           
#>  [52] "b'1(R)"            "ICL(L)"            "FBl1"             
#>  [55] "AVLP(R)"           "VES(L)"            "LOP(R)"           
#>  [58] "CRE(-ROB,-RUB)(R)" "CA(L)"             "a2(R)"            
#>  [61] "SAD"               "LAL(-GA)(R)"       "g2(R)"            
#>  [64] "NO2(R)"            "VLNP(R)"           "lACA(R)"          
#>  [67] "VMNP"              "bL(L)"             "SAD(-AMMC)"       
#>  [70] "FBl8"              "BU(L)"             "FBl5"             
#>  [73] "GC"                "EBr3pw"            "bL(R)"            
#>  [76] "a'3(R)"            "VES(R)"            "AB(R)"            
#>  [79] "LX(L)"             "NO1(L)"            "EBr5"             
#>  [82] "gL(L)"             "ROB(R)"            "PB(L7)"           
#>  [85] "aL(L)"             "WED(R)"            "PB(L1)"           
#>  [88] "IPS(R)"            "EPA(L)"            "POC"              
#>  [91] "hemibrain"         "NO3(R)"            "SPS(L)"           
#>  [94] "SNP(L)"            "PB(R2)"            "PB(L2)"           
#>  [97] "a3(R)"             "b'L(L)"            "NO2(L)"           
#> [100] "PB(R7)"            "b'L(R)"            "SCL(R)"           
#> [103] "BU(R)"             "g5(R)"             "FBl9"             
#> [106] "LH(R)"             "PB(L4)"            "AB(L)"            
#> [109] "SCL(L)"            "EBr1"              "CAN(R)"           
#> [112] "AL(R)"             "mALT(L)"           "g3(R)"            
#> [115] "EPA(R)"            "gL(R)"             "CA(R)"            
#> [118] "IB"                "PB(L3)"            "FB"               
#> [121] "g4(R)"             "RUB(L)"            "a'L(L)"           
#> [124] "OL(R)"             "vACA(R)"           "a1(R)"            
#> [127] "SPS(R)"            "g1(R)"             "a'2(R)"           
#> [130] "CRE(L)"            "PB(R3)"            "GNG"              
#> [133] "dACA(R)"           "EBr3d"             "MB(+ACA)(R)"      
#> [136] "FBl3"              "AME(R)"            "MB(L)"            
#> [139] "SNP(R)"            "EB"                "FLA(R)"           
#> [142] "PB(R9)"            "a'1(R)"            "b'2(R)"           
#> [145] "PB(L8)"            "PB(L9)"            "PB(L5)"           
#> [148] "a'L(R)"            "PB(R1)"            "EBr6"             
#> [151] "NO(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                        
#> [1,] "POST" "/api/v:ver/raw/cypher/transaction/:id/kill"
#> [2,] "GET"  "/api/cached/roicompleteness"               
#> [3,] "GET"  "/api/npexplorer/roisinneuron"              
#> [4,] "POST" "/api/v:ver/npexplorer/completeness"        
#> [5,] "POST" "/api/skeletons/skeleton/:dataset/:id"      
#> [6,] "GET"  "/api/roimeshes/mesh/:dataset/:roi"         
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROICompleteness-fm"    
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [5,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/roimeshes.masterAPI.getMesh-fm"            

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }