Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(R6)"            "LAL(L)"            "g2(R)"            
#>   [4] "PB(L7)"            "b1(R)"             "AOTU(R)"          
#>   [7] "lACA(R)"           "MB(L)"             "EBr6"             
#>  [10] "BU(L)"             "ME(R)"             "PB(L9)"           
#>  [13] "bL(L)"             "FBl9"              "PB(R5)"           
#>  [16] "GOR(L)"            "FLA(R)"            "MB(+ACA)(R)"      
#>  [19] "ICL(R)"            "ATL(L)"            "dACA(R)"          
#>  [22] "PB(L8)"            "g5(R)"             "CRE(R)"           
#>  [25] "PB"                "PLP(R)"            "EB"               
#>  [28] "ICL(L)"            "AB(R)"             "b'2(R)"           
#>  [31] "mALT(R)"           "LX(R)"             "NO"               
#>  [34] "PB(L1)"            "EBr2r4"            "a'L(R)"           
#>  [37] "POC"               "SMP(R)"            "GF(R)"            
#>  [40] "AL(L)"             "SAD(-AMMC)"        "IB"               
#>  [43] "PB(R3)"            "AVLP(R)"           "RUB(R)"           
#>  [46] "PB(R1)"            "PRW"               "NO1(L)"           
#>  [49] "gL(R)"             "a'1(R)"            "FBl5"             
#>  [52] "NO2(R)"            "SCL(L)"            "SPS(R)"           
#>  [55] "g1(R)"             "PB(R8)"            "EBr3d"            
#>  [58] "SNP(R)"            "LO(R)"             "LX(L)"            
#>  [61] "aL(L)"             "GOR(R)"            "MB(R)"            
#>  [64] "SPS(L)"            "PB(R4)"            "NO3(R)"           
#>  [67] "VMNP"              "FBl1"              "GNG"              
#>  [70] "vACA(R)"           "g3(R)"             "NO3(L)"           
#>  [73] "RUB(L)"            "a3(R)"             "LAL(R)"           
#>  [76] "PB(R7)"            "b2(R)"             "INP"              
#>  [79] "CX"                "a2(R)"             "gL(L)"            
#>  [82] "SNP(L)"            "CAN(R)"            "FBl2"             
#>  [85] "LH(R)"             "EPA(L)"            "aL(R)"            
#>  [88] "NO(R)"             "FBl8"              "AB(L)"            
#>  [91] "PB(L5)"            "SCL(R)"            "FBl3"             
#>  [94] "bL(R)"             "NO(L)"             "OL(R)"            
#>  [97] "a'3(R)"            "GA(R)"             "ROB(R)"           
#> [100] "LOP(R)"            "EBr5"              "AMMC"             
#> [103] "PENP"              "WED(R)"            "CA(R)"            
#> [106] "IPS(R)"            "PB(L2)"            "EBr3am"           
#> [109] "FB"                "g4(R)"             "EPA(R)"           
#> [112] "EBr3pw"            "VES(R)"            "PVLP(R)"          
#> [115] "FBl6"              "FBl7"              "mALT(L)"          
#> [118] "CA(L)"             "SIP(L)"            "PED(R)"           
#> [121] "ATL(R)"            "b'L(L)"            "LAL(-GA)(R)"      
#> [124] "GC"                "NO1(R)"            "BU(R)"            
#> [127] "FBl4"              "PB(R2)"            "VES(L)"           
#> [130] "CRE(-ROB,-RUB)(R)" "NO2(L)"            "SMP(L)"           
#> [133] "hemibrain"         "SAD"               "a'2(R)"           
#> [136] "AME(R)"            "AL(R)"             "PB(L3)"           
#> [139] "a'L(L)"            "SLP(R)"            "AOT(R)"           
#> [142] "b'L(R)"            "EBr1"              "PB(L4)"           
#> [145] "VLNP(R)"           "b'1(R)"            "a1(R)"            
#> [148] "PB(R9)"            "PB(L6)"            "SIP(R)"           
#> [151] "CRE(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-VA4(R)"         "FBl8"              "AL-VP3(R)"        
#>   [4] "PB(R6)"            "NO"                "NO(R)"            
#>   [7] "FB-column3"        "AL-DL1(R)"         "EBr5"             
#>  [10] "VMNP"              "NO2(R)"            "AL-DC1(R)"        
#>  [13] "a'1(R)"            "FBl1"              "FBl9"             
#>  [16] "AL-DC4(R)"         "lACA(R)"           "CRE(R)"           
#>  [19] "a'2(R)"            "AL-DL4(L)"         "NO1(L)"           
#>  [22] "LX(R)"             "AL-VP4(R)"         "LAL(R)"           
#>  [25] "AL-DM2(R)"         "PRW"               "AL-VA1d(R)"       
#>  [28] "SCL(L)"            "PB(L8)"            "AL-VC2(R)"        
#>  [31] "PLP(R)"            "AL-VA6(L)"         "FBl4"             
#>  [34] "NO1(R)"            "AL-DL3(R)"         "AL-VP2(R)"        
#>  [37] "AL-VM7d(L)"        "SLP(R)"            "mALT(L)"          
#>  [40] "PENP"              "g1(R)"             "aL(L)"            
#>  [43] "AL-VA2(R)"         "CRE(-ROB,-RUB)(R)" "SAD(-AMMC)"       
#>  [46] "bL(R)"             "PB"                "SIP(R)"           
#>  [49] "AL-VM4(R)"         "AOTU(R)"           "PED(R)"           
#>  [52] "AL-D(L)"           "AL-DM6(R)"         "AVLP(R)"          
#>  [55] "PB(L4)"            "AL-DM4(R)"         "gL(L)"            
#>  [58] "AL-DL2v(R)"        "g3(R)"             "g4(R)"            
#>  [61] "AL-DM6(L)"         "VES(R)"            "AL-DA2(R)"        
#>  [64] "POC"               "PB(L6)"            "LH(R)"            
#>  [67] "AL-VC4(R)"         "GF(R)"             "PB(L3)"           
#>  [70] "hemibrain"         "SAD"               "LX(L)"            
#>  [73] "a3(R)"             "MB(R)"             "AB(L)"            
#>  [76] "FBl6"              "ATL(R)"            "AL-VC1(R)"        
#>  [79] "AL-VP5(R)"         "SMP(R)"            "b'L(L)"           
#>  [82] "AL-DM1(R)"         "AL-D(R)"           "AL-VA7m(R)"       
#>  [85] "AL-DM3(R)"         "AL-VA1v(R)"        "vACA(R)"          
#>  [88] "WED(R)"            "AL-DM1(L)"         "AL-DP1m(L)"       
#>  [91] "PB(R2)"            "VES(L)"            "AL-VA7l(R)"       
#>  [94] "AL-VC5(R)"         "a'3(R)"            "SIP(L)"           
#>  [97] "PB(R7)"            "FBl3"              "AB(R)"            
#> [100] "EBr3am"            "ATL(L)"            "CRE(-RUB)(L)"     
#> [103] "a'L(L)"            "AL(R)"             "AL-VM5v(R)"       
#> [106] "EPA(L)"            "FB"                "PB(L1)"           
#> [109] "AL-V(R)"           "SNP(R)"            "PB(L9)"           
#> [112] "CRE(L)"            "AL-VM3(R)"         "b2(R)"            
#> [115] "AL-DA2(L)"         "LOP(R)"            "MB(+ACA)(R)"      
#> [118] "FBl5"              "AL-DA4l(R)"        "AL-DA4m(R)"       
#> [121] "dACA(R)"           "EB"                "GNG"              
#> [124] "VLNP(R)"           "EBr3d"             "NO(L)"            
#> [127] "EBr6"              "AL(L)"             "AL-DC1(L)"        
#> [130] "FBl2"              "BU(R)"             "AL-DM2(L)"        
#> [133] "SNP(L)"            "AL-VL2p(R)"        "AL-VP1l(R)"       
#> [136] "EBr3pw"            "RUB(R)"            "CAN(R)"           
#> [139] "PB(R1)"            "b1(R)"             "RUB(L)"           
#> [142] "NO2(L)"            "mALT(R)"           "NO3(R)"           
#> [145] "EBr2r4"            "a'L(R)"            "AL-DC2(L)"        
#> [148] "AL-DL5(L)"         "PB(R3)"            "PB(R4)"           
#> [151] "AL-VM7v(R)"        "a1(R)"             "LAL(L)"           
#> [154] "AL-VA3(R)"         "AL-VC3m(R)"        "AL-VP1d(R)"       
#> [157] "AL-DM4(L)"         "AL-VP1m(R)"        "PB(L5)"           
#> [160] "b'2(R)"            "MB(L)"             "BU(L)"            
#> [163] "FLA(R)"            "LAL(-GA)(R)"       "ICL(R)"           
#> [166] "b'L(R)"            "IPS(R)"            "AL-DC2(R)"        
#> [169] "SMP(L)"            "AL-DM3(L)"         "AL-DM5(L)"        
#> [172] "g5(R)"             "g2(R)"             "GOR(R)"           
#> [175] "AL-DM5(R)"         "IB"                "INP"              
#> [178] "a2(R)"             "ME(R)"             "PB(L7)"           
#> [181] "AL-VC3l(R)"        "AOT(R)"            "GC"               
#> [184] "AL-DC4(L)"         "AL-DA3(R)"         "AL-VL1(R)"        
#> [187] "AL-VM2(R)"         "CA(L)"             "SPS(L)"           
#> [190] "AL-DA4m(L)"        "ICL(L)"            "CA(R)"            
#> [193] "GA(R)"             "AL-DA3(L)"         "AL-VL2a(R)"       
#> [196] "PB(R5)"            "PB(L2)"            "OL(R)"            
#> [199] "AL-DA1(R)"         "AL-DL2d(R)"        "AL-DP1l(R)"       
#> [202] "PB(R8)"            "AL-DL5(R)"         "AL-VA6(R)"        
#> [205] "AL-VM1(R)"         "AL-VM5d(R)"        "SPS(R)"           
#> [208] "aL(R)"             "PB(R9)"            "AME(R)"           
#> [211] "EBr1"              "AMMC"              "GOR(L)"           
#> [214] "SCL(R)"            "LO(R)"             "gL(R)"            
#> [217] "AL-VA5(R)"         "AL-VM7d(R)"        "b'1(R)"           
#> [220] "EPA(R)"            "AL-VM7v(L)"        "ROB(R)"           
#> [223] "bL(L)"             "NO3(L)"            "AL-DL4(R)"        
#> [226] "PVLP(R)"           "FBl7"              "AL-DP1m(R)"       
#> [229] "AL-DC3(R)"         "CX"               
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                                
#> [1,] "GET"     "/api/v:ver/custom/custom"          
#> [2,] "POST"    "/api/v:ver/npexplorer/completeness"
#> [3,] "HEAD"    "/api/*"                            
#> [4,] "GET"     "/api/npexplorer/autapses"          
#> [5,] "GET"     "/api/cached/roiconnectivity"       
#> [6,] "OPTIONS" "/api"                              
#>      name                                                                                    
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [3,] "github.com/labstack/echo.glob..func1"                                                  
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROIConnectivity-fm"
#> [6,] "github.com/labstack/echo.glob..func1"                                                  

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }