Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "NO3(L)"            "IPS(R)"            "CRE(-ROB,-RUB)(R)"
#>   [4] "VES(R)"            "AMMC"              "SAD"              
#>   [7] "EBr6"              "FLA(R)"            "ROB(R)"           
#>  [10] "MB(R)"             "FBl6"              "LAL(-GA)(R)"      
#>  [13] "BU(L)"             "NO(R)"             "SPS(L)"           
#>  [16] "SNP(R)"            "a'L(R)"            "LX(R)"            
#>  [19] "a3(R)"             "PB(L5)"            "SPS(R)"           
#>  [22] "mALT(R)"           "b'L(R)"            "FBl2"             
#>  [25] "GNG"               "PB(L4)"            "b'1(R)"           
#>  [28] "IB"                "b'2(R)"            "NO1(L)"           
#>  [31] "PB(R1)"            "EBr1"              "g4(R)"            
#>  [34] "ME(R)"             "PB(R3)"            "AL(R)"            
#>  [37] "RUB(R)"            "a'3(R)"            "GF(R)"            
#>  [40] "PB(L3)"            "PB(L1)"            "BU(R)"            
#>  [43] "b'L(L)"            "AOT(R)"            "g1(R)"            
#>  [46] "a'1(R)"            "NO2(L)"            "EBr2r4"           
#>  [49] "SNP(L)"            "PB(L2)"            "ATL(L)"           
#>  [52] "CA(L)"             "PLP(R)"            "SLP(R)"           
#>  [55] "AB(L)"             "NO3(R)"            "FB"               
#>  [58] "bL(R)"             "SIP(R)"            "EBr3pw"           
#>  [61] "SCL(L)"            "SMP(L)"            "a2(R)"            
#>  [64] "WED(R)"            "FBl7"              "PB(R2)"           
#>  [67] "AVLP(R)"           "EPA(L)"            "NO(L)"            
#>  [70] "EPA(R)"            "PB(R9)"            "FBl3"             
#>  [73] "PB(R6)"            "LX(L)"             "aL(R)"            
#>  [76] "PB"                "PVLP(R)"           "FBl4"             
#>  [79] "FBl8"              "AB(R)"             "aL(L)"            
#>  [82] "PB(R8)"            "NO"                "PB(L6)"           
#>  [85] "FBl9"              "GOR(R)"            "VES(L)"           
#>  [88] "PB(R7)"            "g2(R)"             "EBr5"             
#>  [91] "OL(R)"             "SMP(R)"            "a1(R)"            
#>  [94] "LAL(L)"            "NO1(R)"            "ICL(R)"           
#>  [97] "b1(R)"             "POC"               "dACA(R)"          
#> [100] "AL(L)"             "SAD(-AMMC)"        "LAL(R)"           
#> [103] "AOTU(R)"           "EBr3d"             "EB"               
#> [106] "MB(L)"             "SCL(R)"            "PB(L8)"           
#> [109] "CA(R)"             "g3(R)"             "GOR(L)"           
#> [112] "GC"                "SIP(L)"            "LH(R)"            
#> [115] "EBr3am"            "PED(R)"            "CRE(R)"           
#> [118] "MB(+ACA)(R)"       "PB(R5)"            "mALT(L)"          
#> [121] "bL(L)"             "CAN(R)"            "INP"              
#> [124] "ICL(L)"            "FBl5"              "VLNP(R)"          
#> [127] "hemibrain"         "lACA(R)"           "a'2(R)"           
#> [130] "vACA(R)"           "PB(L7)"            "ATL(R)"           
#> [133] "a'L(L)"            "CRE(L)"            "g5(R)"            
#> [136] "FBl1"              "RUB(L)"            "PB(R4)"           
#> [139] "VMNP"              "gL(R)"             "PB(L9)"           
#> [142] "LOP(R)"            "LO(R)"             "NO2(R)"           
#> [145] "PENP"              "GA(R)"             "PRW"              
#> [148] "CX"                "b2(R)"             "gL(L)"            
#> [151] "AME(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                    
#> [1,] "POST" "/api/raw/cypher/transaction/:id/cypher"
#> [2,] "GET"  "/api/v:ver/cached/dailytype"           
#> [3,] "GET"  "/api/v:ver/dbmeta/datasets"            
#> [4,] "POST" "/api/v:ver/npexplorer/rankedtable"     
#> [5,] "POST" "/api/raw/cypher/transaction"           
#> [6,] "HEAD" "/api/*"                                
#>      name                                                                    
#> [1,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"    
#> [5,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"
#> [6,] "github.com/labstack/echo.glob..func1"                                  

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }