Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "RUB(R)"            "PB(L5)"            "mALT(R)"          
#>   [4] "g5(R)"             "LX(L)"             "INP"              
#>   [7] "FBl5"              "ICL(R)"            "aL(L)"            
#>  [10] "a3(R)"             "EBr1"              "LX(R)"            
#>  [13] "EB"                "BU(R)"             "ATL(L)"           
#>  [16] "AMMC"              "FB"                "g2(R)"            
#>  [19] "SNP(L)"            "FBl2"              "PB(R6)"           
#>  [22] "b'2(R)"            "LAL(L)"            "SAD"              
#>  [25] "FBl4"              "CA(R)"             "EBr3am"           
#>  [28] "a'3(R)"            "a1(R)"             "PED(R)"           
#>  [31] "gL(R)"             "FBl7"              "LAL(R)"           
#>  [34] "IPS(R)"            "MB(R)"             "WED(R)"           
#>  [37] "b2(R)"             "VES(R)"            "FBl1"             
#>  [40] "GOR(R)"            "bL(R)"             "IB"               
#>  [43] "NO3(R)"            "PB(L1)"            "EBr2r4"           
#>  [46] "GNG"               "PRW"               "a'1(R)"           
#>  [49] "AB(R)"             "b'1(R)"            "VMNP"             
#>  [52] "LH(R)"             "PB(R3)"            "SCL(L)"           
#>  [55] "vACA(R)"           "lACA(R)"           "aL(R)"            
#>  [58] "MB(+ACA)(R)"       "FBl3"              "NO3(L)"           
#>  [61] "AVLP(R)"           "AL(R)"             "AOTU(R)"          
#>  [64] "GOR(L)"            "MB(L)"             "PB(R9)"           
#>  [67] "a'2(R)"            "EBr5"              "b1(R)"            
#>  [70] "GF(R)"             "PB(L3)"            "dACA(R)"          
#>  [73] "FBl9"              "CA(L)"             "g3(R)"            
#>  [76] "SPS(R)"            "LAL(-GA)(R)"       "NO1(L)"           
#>  [79] "GC"                "CRE(L)"            "bL(L)"            
#>  [82] "PB(L8)"            "VLNP(R)"           "SLP(R)"           
#>  [85] "gL(L)"             "EPA(R)"            "PB(L9)"           
#>  [88] "PB(R2)"            "FBl8"              "AME(R)"           
#>  [91] "mALT(L)"           "hemibrain"         "SIP(L)"           
#>  [94] "PB(R7)"            "PB(L6)"            "NO"               
#>  [97] "ROB(R)"            "ATL(R)"            "CAN(R)"           
#> [100] "NO(L)"             "SMP(L)"            "ME(R)"            
#> [103] "SIP(R)"            "LO(R)"             "a'L(R)"           
#> [106] "SAD(-AMMC)"        "RUB(L)"            "CRE(-ROB,-RUB)(R)"
#> [109] "PB(R1)"            "PB(L7)"            "a'L(L)"           
#> [112] "AB(L)"             "PB(L2)"            "CRE(R)"           
#> [115] "NO1(R)"            "NO(R)"             "EPA(L)"           
#> [118] "SCL(R)"            "EBr6"              "OL(R)"            
#> [121] "SMP(R)"            "VES(L)"            "POC"              
#> [124] "PB(R5)"            "LOP(R)"            "CX"               
#> [127] "NO2(L)"            "b'L(R)"            "PB"               
#> [130] "EBr3d"             "b'L(L)"            "GA(R)"            
#> [133] "SPS(L)"            "PB(L4)"            "EBr3pw"           
#> [136] "FBl6"              "g1(R)"             "FLA(R)"           
#> [139] "PLP(R)"            "SNP(R)"            "BU(L)"            
#> [142] "NO2(R)"            "PB(R4)"            "a2(R)"            
#> [145] "g4(R)"             "PVLP(R)"           "AOT(R)"           
#> [148] "PENP"              "PB(R8)"            "AL(L)"            
#> [151] "ICL(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                               
#> [1,] "PUT"     "/api"                             
#> [2,] "GET"     "/api/v:ver/custom/custom"         
#> [3,] "POST"    "/api/npexplorer/findneurons"      
#> [4,] "POST"    "/api/npexplorer/simpleconnections"
#> [5,] "GET"     "/api/npexplorer/completeness"     
#> [6,] "CONNECT" "/api"                             
#>      name                                                                                             
#> [1,] "github.com/labstack/echo.glob..func1"                                                           
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getFindNeurons-fm"      
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCompleteness-fm"     
#> [6,] "github.com/labstack/echo.glob..func1"                                                           

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }