Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $fib19
#> $fib19$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $fib19$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $fib19$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $fib19$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $fib19$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $fib19$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "lACA(R)"           "EB"                "AL-DM4(R)"        
#>   [4] "AL-VL1(R)"         "AL-VM7v(R)"        "GOR(L)"           
#>   [7] "a'2(R)"            "LO(R)"             "CA(R)"            
#>  [10] "AL-DA1(R)"         "AL-VC5(R)"         "AL(R)"            
#>  [13] "PRW"               "NO(R)"             "GNG"              
#>  [16] "b'L(L)"            "NO(L)"             "AL-VP4(R)"        
#>  [19] "g5(R)"             "AL-DC3(R)"         "PENP"             
#>  [22] "LAL(L)"            "PB(R7)"            "EBr1"             
#>  [25] "EBr6"              "ICL(L)"            "EPA(R)"           
#>  [28] "FBl7"              "MB(L)"             "VES(R)"           
#>  [31] "CAN(R)"            "AL-DL5(R)"         "AL-DM5(L)"        
#>  [34] "AL-DM5(R)"         "PED(R)"            "AL-DM1(L)"        
#>  [37] "GOR(R)"            "AL-VC4(R)"         "AL-VM1(R)"        
#>  [40] "PB(R6)"            "AL-VL2p(R)"        "hemibrain"        
#>  [43] "WED(R)"            "aL(R)"             "BU(L)"            
#>  [46] "SPS(R)"            "EBr3d"             "AL-DA4l(R)"       
#>  [49] "AL-DM1(R)"         "AL-DM2(R)"         "AL-VA3(R)"        
#>  [52] "FBl9"              "AME(R)"            "AL-DA4m(R)"       
#>  [55] "PB(R5)"            "EBr5"              "CRE(L)"           
#>  [58] "CRE(-RUB)(L)"      "FBl8"              "PB(L1)"           
#>  [61] "NO2(R)"            "PB"                "AL-DC2(R)"        
#>  [64] "AL-VC3l(R)"        "PB(R8)"            "AL-DM3(L)"        
#>  [67] "a'3(R)"            "ROB(R)"            "FB-column3"       
#>  [70] "FBl5"              "a'L(L)"            "AL-VP2(R)"        
#>  [73] "b2(R)"             "PB(L2)"            "PB(L6)"           
#>  [76] "AL-DP1m(R)"        "AL-VP5(R)"         "AL-VC1(R)"        
#>  [79] "AOTU(R)"           "PB(R9)"            "RUB(L)"           
#>  [82] "SLP(R)"            "AL-VA1d(R)"        "AL-VA2(R)"        
#>  [85] "NO1(R)"            "SIP(R)"            "NO"               
#>  [88] "FBl1"              "AL-VM4(R)"         "g3(R)"            
#>  [91] "NO3(L)"            "EPA(L)"            "AL-VP1l(R)"       
#>  [94] "IB"                "CRE(R)"            "AL-DM4(L)"        
#>  [97] "PB(R1)"            "INP"               "AL-DL2v(R)"       
#> [100] "POC"               "a2(R)"             "g4(R)"            
#> [103] "PB(L4)"            "NO3(R)"            "AL-DC2(L)"        
#> [106] "AL-VC3m(R)"        "b'2(R)"            "AL-DA3(L)"        
#> [109] "bL(R)"             "AL-VA5(R)"         "AL-VC2(R)"        
#> [112] "AL-DM2(L)"         "PB(L9)"            "MB(+ACA)(R)"      
#> [115] "RUB(R)"            "mALT(L)"           "SMP(L)"           
#> [118] "PVLP(R)"           "AL-D(R)"           "AL-DC1(L)"        
#> [121] "AOT(R)"            "CX"                "SAD"              
#> [124] "VMNP"              "IB112"             "AVLP(R)"          
#> [127] "PB(L5)"            "AL-VA7l(R)"        "LAL(-GA)(R)"      
#> [130] "NO1(L)"            "PB(L8)"            "FBl6"             
#> [133] "AL-DM3(R)"         "AL-DM6(R)"         "a1(R)"            
#> [136] "AL-DA2(L)"         "AL-VM7v(L)"        "LAL(R)"           
#> [139] "AL-VA4(R)"         "AL-VM7d(R)"        "b'L(R)"           
#> [142] "EBr2r4"            "MB(R)"             "AB(L)"            
#> [145] "AL-DA4m(L)"        "AL-DP1m(L)"        "AL-DC1(R)"        
#> [148] "GC"                "FLA(R)"            "SNP(R)"           
#> [151] "AL-VA7m(R)"        "AL-VP1d(R)"        "AL-D(L)"          
#> [154] "bL(L)"             "LOP(R)"            "FBl4"             
#> [157] "b1(R)"             "AL-DM6(L)"         "AL-DL2d(R)"       
#> [160] "AL-V(R)"           "ATL(L)"            "CA(L)"            
#> [163] "gL(L)"             "AL(L)"             "AL-DL5(L)"        
#> [166] "NO2(L)"            "VLNP(R)"           "mALT(R)"          
#> [169] "ICL(R)"            "LX(L)"             "AL-DA3(R)"        
#> [172] "AL-VM5v(R)"        "AL-VP1m(R)"        "ATL(R)"           
#> [175] "SCL(R)"            "AL-VM3(R)"         "EBr3pw"           
#> [178] "CRE(-ROB,-RUB)(R)" "SMP(R)"            "AL-DL4(L)"        
#> [181] "SAD(-AMMC)"        "FBl3"              "LH(R)"            
#> [184] "AL-VA6(R)"         "AL-VM2(R)"         "FB"               
#> [187] "AL-VM5d(R)"        "SCL(L)"            "EBr3am"           
#> [190] "PB(L3)"            "AL-DP1l(R)"        "GF(R)"            
#> [193] "g2(R)"             "AL-DC4(L)"         "SNP(L)"           
#> [196] "AL-DL1(R)"         "dACA(R)"           "PB(L7)"           
#> [199] "AL-VA1v(R)"        "LX(R)"             "a3(R)"            
#> [202] "AL-DA2(R)"         "AL-VP3(R)"         "AMMC"             
#> [205] "SIP(L)"            "g1(R)"             "a'L(R)"           
#> [208] "BU(R)"             "a'1(R)"            "VES(L)"           
#> [211] "ME(R)"             "PB(R2)"            "AB(R)"            
#> [214] "AL-DC4(R)"         "AL-DL3(R)"         "PLP(R)"           
#> [217] "AL-VA6(L)"         "PB(R3)"            "PB(R4)"           
#> [220] "AL-DL4(R)"         "vACA(R)"           "GA(R)"            
#> [223] "gL(R)"             "aL(L)"             "OL(R)"            
#> [226] "SPS(L)"            "IPS(R)"            "AL-VM7d(L)"       
#> [229] "FBl2"              "AL-VL2a(R)"        "b'1(R)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "LX(R)"             "BU(R)"             "NO3(L)"           
#>   [4] "CRE(R)"            "bL(L)"             "VES(R)"           
#>   [7] "LH(R)"             "AL(R)"             "NO1(R)"           
#>  [10] "FBl8"              "AB(R)"             "GA(R)"            
#>  [13] "CA(L)"             "SIP(R)"            "PB(L5)"           
#>  [16] "FBl2"              "g1(R)"             "EBr1"             
#>  [19] "gL(R)"             "CA(R)"             "LAL(R)"           
#>  [22] "NO2(L)"            "WED(R)"            "PB(L1)"           
#>  [25] "PB(R8)"            "EBr3d"             "IPS(R)"           
#>  [28] "lACA(R)"           "FB"                "PB"               
#>  [31] "PB(R2)"            "AVLP(R)"           "aL(L)"            
#>  [34] "CRE(-ROB,-RUB)(R)" "ATL(L)"            "a2(R)"            
#>  [37] "AL(L)"             "NO"                "bL(R)"            
#>  [40] "SLP(R)"            "a'3(R)"            "b2(R)"            
#>  [43] "g3(R)"             "GOR(L)"            "EB"               
#>  [46] "CRE(L)"            "b'L(L)"            "LAL(L)"           
#>  [49] "MB(+ACA)(R)"       "FBl9"              "a3(R)"            
#>  [52] "LOP(R)"            "mALT(L)"           "FLA(R)"           
#>  [55] "SPS(L)"            "a'L(R)"            "ROB(R)"           
#>  [58] "ICL(L)"            "PB(L7)"            "LX(L)"            
#>  [61] "CAN(R)"            "BU(L)"             "a'L(L)"           
#>  [64] "b'1(R)"            "GNG"               "PVLP(R)"          
#>  [67] "AME(R)"            "gL(L)"             "EBr5"             
#>  [70] "EBr2r4"            "NO2(R)"            "hemibrain"        
#>  [73] "FBl6"              "NO(L)"             "RUB(R)"           
#>  [76] "IB"                "MB(L)"             "PB(L8)"           
#>  [79] "ICL(R)"            "EBr6"              "ME(R)"            
#>  [82] "EPA(R)"            "PB(L6)"            "FBl1"             
#>  [85] "FBl7"              "RUB(L)"            "AB(L)"            
#>  [88] "b'2(R)"            "PB(R3)"            "EPA(L)"           
#>  [91] "PB(R4)"            "VES(L)"            "g4(R)"            
#>  [94] "SAD"               "LAL(-GA)(R)"       "NO1(L)"           
#>  [97] "g2(R)"             "PB(R9)"            "SNP(L)"           
#> [100] "AOTU(R)"           "FBl3"              "mALT(R)"          
#> [103] "SPS(R)"            "AMMC"              "a1(R)"            
#> [106] "b1(R)"             "PB(L4)"            "SAD(-AMMC)"       
#> [109] "VLNP(R)"           "MB(R)"             "dACA(R)"          
#> [112] "PB(R1)"            "PB(L9)"            "SCL(L)"           
#> [115] "SIP(L)"            "PLP(R)"            "EBr3pw"           
#> [118] "PENP"              "aL(R)"             "b'L(R)"           
#> [121] "FBl5"              "PB(R5)"            "PRW"              
#> [124] "GC"                "a'2(R)"            "PB(R6)"           
#> [127] "PB(R7)"            "GOR(R)"            "AOT(R)"           
#> [130] "SMP(R)"            "GF(R)"             "PED(R)"           
#> [133] "SCL(R)"            "VMNP"              "g5(R)"            
#> [136] "INP"               "OL(R)"             "POC"              
#> [139] "PB(L3)"            "FBl4"              "ATL(R)"           
#> [142] "PB(L2)"            "SNP(R)"            "NO3(R)"           
#> [145] "NO(R)"             "a'1(R)"            "LO(R)"            
#> [148] "EBr3am"            "vACA(R)"           "CX"               
#> [151] "SMP(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "FBl4"              "b'L(L)"            "bL(L)"            
#>   [4] "SPS(L)"            "MB(+ACA)(R)"       "a'L(L)"           
#>   [7] "AL-DL4(R)"         "AL-VL2p(R)"        "SCL(R)"           
#>  [10] "AL-DM4(L)"         "AL-D(L)"           "AL-DC3(R)"        
#>  [13] "AL-DM5(R)"         "AL-VC3l(R)"        "FB"               
#>  [16] "MB(R)"             "AL-VC4(R)"         "PB(L5)"           
#>  [19] "g5(R)"             "ATL(L)"            "g3(R)"            
#>  [22] "AL(L)"             "AL-DL5(R)"         "PRW"              
#>  [25] "AL-VM7v(L)"        "RUB(L)"            "PB(L3)"           
#>  [28] "a'L(R)"            "gL(L)"             "PB(R9)"           
#>  [31] "AL-DP1l(R)"        "GC"                "LOP(R)"           
#>  [34] "SLP(R)"            "FBl2"              "a3(R)"            
#>  [37] "EBr3pw"            "EBr3d"             "AL-DA3(L)"        
#>  [40] "CAN(R)"            "AL-DC1(R)"         "dACA(R)"          
#>  [43] "ATL(R)"            "AL-VC2(R)"         "BU(L)"            
#>  [46] "RUB(R)"            "AL-DM1(L)"         "a'1(R)"           
#>  [49] "AL-DP1m(R)"        "AL-VM5d(R)"        "PED(R)"           
#>  [52] "b2(R)"             "FBl8"              "LH(R)"            
#>  [55] "GF(R)"             "SNP(R)"            "AL-DC4(L)"        
#>  [58] "ICL(L)"            "AL-VA7m(R)"        "g1(R)"            
#>  [61] "AL-VL2a(R)"        "AL-DM2(L)"         "bL(R)"            
#>  [64] "AL-DL1(R)"         "ROB(R)"            "FBl6"             
#>  [67] "AL-VA6(L)"         "AVLP(R)"           "NO(L)"            
#>  [70] "AL-VA1d(R)"        "AL-VA4(R)"         "NO3(R)"           
#>  [73] "BU(R)"             "AL-DM3(L)"         "PB(R2)"           
#>  [76] "AL-DA2(R)"         "AL-VM7d(R)"        "SCL(L)"           
#>  [79] "PB"                "AL-DM4(R)"         "PB(L9)"           
#>  [82] "mALT(L)"           "PVLP(R)"           "AL-DA1(R)"        
#>  [85] "ME(R)"             "AL-VM7d(L)"        "AL-VA7l(R)"       
#>  [88] "EBr3am"            "CX"                "hemibrain"        
#>  [91] "PB(L4)"            "PB(R3)"            "NO2(R)"           
#>  [94] "VES(L)"            "VMNP"              "WED(R)"           
#>  [97] "NO1(L)"            "AL-DA2(L)"         "SNP(L)"           
#> [100] "AME(R)"            "AL-VA5(R)"         "PB(R5)"           
#> [103] "SAD"               "AL-VP1l(R)"        "PB(L1)"           
#> [106] "FLA(R)"            "NO(R)"             "AL-DM2(R)"        
#> [109] "PB(R6)"            "PB(L8)"            "SAD(-AMMC)"       
#> [112] "EPA(L)"            "AL-DL3(R)"         "AL-VC5(R)"        
#> [115] "AMMC"              "aL(R)"             "LO(R)"            
#> [118] "AL-DM5(L)"         "NO"                "FBl1"             
#> [121] "AL-VM3(R)"         "b'2(R)"            "LX(L)"            
#> [124] "AL-VM4(R)"         "AL-VP3(R)"         "SPS(R)"           
#> [127] "AL-VC1(R)"         "PENP"              "b1(R)"            
#> [130] "CRE(-RUB)(L)"      "MB(L)"             "PB(R8)"           
#> [133] "AL-DL4(L)"         "AL-DP1m(L)"        "SIP(R)"           
#> [136] "CRE(-ROB,-RUB)(R)" "GA(R)"             "EBr6"             
#> [139] "AB(R)"             "CRE(R)"            "gL(R)"            
#> [142] "AL-VP1d(R)"        "AL-DL5(L)"         "AL-DM1(R)"        
#> [145] "AL-VL1(R)"         "AL-DL2d(R)"        "AL-VP1m(R)"       
#> [148] "mALT(R)"           "EBr5"              "AL-DM6(L)"        
#> [151] "PB(R1)"            "LAL(R)"            "AL-DC2(R)"        
#> [154] "AL-VM2(R)"         "GOR(L)"            "AB(L)"            
#> [157] "FBl3"              "AL-DC4(R)"         "SMP(L)"           
#> [160] "b'L(R)"            "PLP(R)"            "AL-VA1v(R)"       
#> [163] "a1(R)"             "LAL(L)"            "EB"               
#> [166] "AL-DA4m(R)"        "AL-DL2v(R)"        "LX(R)"            
#> [169] "AL(R)"             "AL-VA3(R)"         "PB(R4)"           
#> [172] "AL-VA2(R)"         "AL-VC3m(R)"        "PB(L2)"           
#> [175] "GNG"               "AL-DA4l(R)"        "AL-VA6(R)"        
#> [178] "aL(L)"             "AL-VP2(R)"         "b'1(R)"           
#> [181] "a2(R)"             "FBl7"              "GOR(R)"           
#> [184] "AL-VP5(R)"         "LAL(-GA)(R)"       "g2(R)"            
#> [187] "EPA(R)"            "OL(R)"             "AL-DA3(R)"        
#> [190] "AL-VM1(R)"         "PB(R7)"            "AL-VM7v(R)"       
#> [193] "AOTU(R)"           "FBl5"              "ICL(R)"           
#> [196] "PB(L7)"            "CRE(L)"            "IPS(R)"           
#> [199] "AL-DM6(R)"         "lACA(R)"           "PB(L6)"           
#> [202] "AL-DM3(R)"         "VLNP(R)"           "POC"              
#> [205] "NO2(L)"            "a'2(R)"            "EBr2r4"           
#> [208] "CA(R)"             "NO1(R)"            "FB-column3"       
#> [211] "INP"               "AL-VM5v(R)"        "SMP(R)"           
#> [214] "a'3(R)"            "SIP(L)"            "VES(R)"           
#> [217] "g4(R)"             "AL-V(R)"           "IB"               
#> [220] "NO3(L)"            "AL-D(R)"           "AL-VP4(R)"        
#> [223] "AOT(R)"            "CA(L)"             "AL-DC2(L)"        
#> [226] "AL-DA4m(L)"        "AL-DC1(L)"         "FBl9"             
#> [229] "vACA(R)"           "EBr1"             
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "FB"                "PLP(R)"            "ME(R)"            
#>   [4] "AL-D(L)"           "NO3(L)"            "FBl7"             
#>   [7] "SIP(L)"            "FBl4"              "AL-DM3(L)"        
#>  [10] "AL-DP1l(R)"        "AL-VA6(L)"         "g4(R)"            
#>  [13] "NO1(R)"            "AL-VM7d(L)"        "FB-column3"       
#>  [16] "AL-VL2p(R)"        "a'L(R)"            "CA(R)"            
#>  [19] "MB(R)"             "AL-VA1v(R)"        "AL-VA7m(R)"       
#>  [22] "PB(L7)"            "AL-VM7v(R)"        "AL-DL5(L)"        
#>  [25] "AL-DM2(L)"         "EBr1"              "AL-DC4(L)"        
#>  [28] "AL-DM2(R)"         "a1(R)"             "LAL(-GA)(R)"      
#>  [31] "g2(R)"             "IB"                "lACA(R)"          
#>  [34] "EBr6"              "AL-DA2(R)"         "AL-DL1(R)"        
#>  [37] "AL-VL2a(R)"        "LOP(R)"            "LAL(L)"           
#>  [40] "AL-DA1(R)"         "NO2(L)"            "EPA(R)"           
#>  [43] "SNP(L)"            "PB(R2)"            "a'1(R)"           
#>  [46] "PB"                "AL-DA2(L)"         "PB(L6)"           
#>  [49] "AVLP(R)"           "NO2(R)"            "AL-DC1(R)"        
#>  [52] "AL-VP4(R)"         "SPS(L)"            "AME(R)"           
#>  [55] "AL-DP1m(R)"        "vACA(R)"           "a2(R)"            
#>  [58] "AL-DM5(R)"         "GOR(R)"            "FBl5"             
#>  [61] "FBl3"              "aL(L)"             "PB(L5)"           
#>  [64] "AL-VM7v(L)"        "AL-DM1(L)"         "FBl9"             
#>  [67] "LX(L)"             "AL-VA5(R)"         "AL-VP3(R)"        
#>  [70] "AL-VP2(R)"         "SCL(L)"            "dACA(R)"          
#>  [73] "SIP(R)"            "b1(R)"             "AL-DM1(R)"        
#>  [76] "PRW"               "EPA(L)"            "AL-DC4(R)"        
#>  [79] "VLNP(R)"           "g5(R)"             "FBl6"             
#>  [82] "AL-VP5(R)"         "EBr3am"            "PVLP(R)"          
#>  [85] "CRE(-RUB)(L)"      "AL-VA1d(R)"        "PB(R8)"           
#>  [88] "VES(R)"            "SNP(R)"            "SMP(L)"           
#>  [91] "BU(L)"             "LX(R)"             "AL(L)"            
#>  [94] "OL(R)"             "b'2(R)"            "GNG"              
#>  [97] "AL-VC5(R)"         "ATL(R)"            "AL-DC1(L)"        
#> [100] "PB(R7)"            "SAD(-AMMC)"        "AL-VC3m(R)"       
#> [103] "PB(R5)"            "AL-DA4m(L)"        "EBr3d"            
#> [106] "PB(R9)"            "FBl1"              "IPS(R)"           
#> [109] "AL-VM5d(R)"        "AL-VM7d(R)"        "aL(R)"            
#> [112] "CRE(L)"            "AL-DM4(R)"         "AOTU(R)"          
#> [115] "GA(R)"             "SCL(R)"            "a'2(R)"           
#> [118] "AL-DP1m(L)"        "PB(L3)"            "AL-DC2(L)"        
#> [121] "PB(R4)"            "AL-DM6(R)"         "AL-VA4(R)"        
#> [124] "AL-VM1(R)"         "mALT(R)"           "bL(R)"            
#> [127] "SLP(R)"            "GOR(L)"            "AL-VP1d(R)"       
#> [130] "EB"                "AL-DM3(R)"         "b'1(R)"           
#> [133] "PB(R6)"            "AL-DM4(L)"         "AL-DM6(L)"        
#> [136] "FBl2"              "PED(R)"            "gL(R)"            
#> [139] "LAL(R)"            "AL-DC3(R)"         "AL-VA2(R)"        
#> [142] "AL-VC4(R)"         "SPS(R)"            "AB(R)"            
#> [145] "b'L(L)"            "PB(L2)"            "GC"               
#> [148] "AL-DM5(L)"         "ROB(R)"            "AL-DL2d(R)"       
#> [151] "ATL(L)"            "GF(R)"             "PB(L8)"           
#> [154] "PB(L9)"            "bL(L)"             "CAN(R)"           
#> [157] "PB(L4)"            "AL-VA3(R)"         "b'L(R)"           
#> [160] "FBl8"              "AL-DA4l(R)"        "a'3(R)"           
#> [163] "b2(R)"             "AL-VP1m(R)"        "AL-DL4(L)"        
#> [166] "FLA(R)"            "EBr3pw"            "g3(R)"            
#> [169] "a3(R)"             "AL-DA3(R)"         "AL-VA7l(R)"       
#> [172] "CRE(R)"            "LH(R)"             "AL-VP1l(R)"       
#> [175] "hemibrain"         "ICL(L)"            "RUB(L)"           
#> [178] "AL(R)"             "AL-DA4m(R)"        "PENP"             
#> [181] "mALT(L)"           "LO(R)"             "AL-VL1(R)"        
#> [184] "AL-VM2(R)"         "AL-VM3(R)"         "CA(L)"            
#> [187] "AL-DC2(R)"         "MB(+ACA)(R)"       "PB(L1)"           
#> [190] "ICL(R)"            "AL-D(R)"           "AB(L)"            
#> [193] "NO(R)"             "PB(R1)"            "AL-VA6(R)"        
#> [196] "AMMC"              "NO"                "AL-VC2(R)"        
#> [199] "MB(L)"             "PB(R3)"            "AL-VC1(R)"        
#> [202] "NO3(R)"            "VES(L)"            "AL-VM5v(R)"       
#> [205] "NO1(L)"            "gL(L)"             "INP"              
#> [208] "CRE(-ROB,-RUB)(R)" "CX"                "EBr2r4"           
#> [211] "NO(L)"             "RUB(R)"            "g1(R)"            
#> [214] "AL-DL2v(R)"        "a'L(L)"            "VMNP"             
#> [217] "AL-DL4(R)"         "AL-DL5(R)"         "AL-DA3(L)"        
#> [220] "AL-DL3(R)"         "POC"               "BU(R)"            
#> [223] "EBr5"              "AL-VC3l(R)"        "AL-VM4(R)"        
#> [226] "SMP(R)"            "SAD"               "AL-V(R)"          
#> [229] "AOT(R)"            "WED(R)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method     path                               
#> [1,] "GET"      "/api/serverinfo"                  
#> [2,] "GET"      "/api/npexplorer/nglayers*"        
#> [3,] "GET"      "/api/v:ver/dbmeta/instances"      
#> [4,] "POST"     "/api/npexplorer/simpleconnections"
#> [5,] "GET"      "/api/version"                     
#> [6,] "PROPFIND" "/api/*"                           
#>      name                                                                                             
#> [1,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                         
#> [2,] "github.com/labstack/echo/v4.StaticDirectoryHandler.func1"                                       
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.(*ConnectomeAPI).getAPIVersion-fm"              
#> [6,] "github.com/labstack/echo/v4.glob..func1"                                                        
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }