Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0`
#> $`hemibrain:v1.0`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0`$ROIs
#>   [1] "a2(R)"             "EBr6"              "b1(R)"            
#>   [4] "EPA(L)"            "SNP(L)"            "aL(L)"            
#>   [7] "PB"                "PLP(R)"            "VMNP"             
#>  [10] "SIP(R)"            "SMP(L)"            "PB(L7)"           
#>  [13] "VES(R)"            "LOP(R)"            "b2(R)"            
#>  [16] "AL(L)"             "SAD(-AMMC)"        "vACA(R)"          
#>  [19] "AMMC"              "b'L(R)"            "PB(R1)"           
#>  [22] "PVLP(R)"           "b'L(L)"            "POC"              
#>  [25] "SCL(L)"            "mALT(L)"           "FBl1"             
#>  [28] "FBl6"              "AME(R)"            "MB(R)"            
#>  [31] "gL(L)"             "ROB(R)"            "a3(R)"            
#>  [34] "PENP"              "PB(L3)"            "GOR(L)"           
#>  [37] "LAL(-GA)(R)"       "GOR(R)"            "VLNP(R)"          
#>  [40] "g2(R)"             "PB(L1)"            "MB(+ACA)(R)"      
#>  [43] "CAN(R)"            "NO3(L)"            "LAL(R)"           
#>  [46] "SLP(R)"            "SMP(R)"            "PB(R2)"           
#>  [49] "INP"               "FBl2"              "FBl4"             
#>  [52] "IB"                "FLA(R)"            "PB(L8)"           
#>  [55] "a'2(R)"            "a'L(R)"            "PB(R5)"           
#>  [58] "mALT(R)"           "g1(R)"             "NO"               
#>  [61] "RUB(L)"            "CRE(L)"            "VES(L)"           
#>  [64] "RUB(R)"            "NO2(R)"            "a'3(R)"           
#>  [67] "PED(R)"            "EBr1"              "EB"               
#>  [70] "BU(R)"             "CRE(R)"            "PB(R8)"           
#>  [73] "LO(R)"             "PB(R7)"            "MB(L)"            
#>  [76] "AOT(R)"            "a1(R)"             "PRW"              
#>  [79] "EBr5"              "EBr3pw"            "CA(R)"            
#>  [82] "PB(L6)"            "g3(R)"             "NO3(R)"           
#>  [85] "dACA(R)"           "NO1(R)"            "AOTU(R)"          
#>  [88] "g4(R)"             "AVLP(R)"           "NO(L)"            
#>  [91] "PB(L5)"            "CX"                "EBr2r4"           
#>  [94] "FBl8"              "b'1(R)"            "ATL(L)"           
#>  [97] "PB(R6)"            "LX(L)"             "PB(L2)"           
#> [100] "EBr3am"            "SCL(R)"            "EBr3d"            
#> [103] "g5(R)"             "CRE(-ROB,-RUB)(R)" "CA(L)"            
#> [106] "aL(R)"             "ICL(L)"            "FBl3"             
#> [109] "GNG"               "LX(R)"             "EPA(R)"           
#> [112] "bL(L)"             "AL(R)"             "b'2(R)"           
#> [115] "hemibrain"         "PB(R9)"            "a'1(R)"           
#> [118] "NO(R)"             "ICL(R)"            "a'L(L)"           
#> [121] "OL(R)"             "SAD"               "SPS(R)"           
#> [124] "GC"                "ME(R)"             "AB(L)"            
#> [127] "PB(L4)"            "GF(R)"             "LH(R)"            
#> [130] "ATL(R)"            "PB(R3)"            "NO1(L)"           
#> [133] "BU(L)"             "gL(R)"             "IPS(R)"           
#> [136] "LAL(L)"            "GA(R)"             "SIP(L)"           
#> [139] "lACA(R)"           "WED(R)"            "bL(R)"            
#> [142] "NO2(L)"            "FB"                "SPS(L)"           
#> [145] "SNP(R)"            "FBl7"              "PB(R4)"           
#> [148] "PB(L9)"            "FBl5"              "FBl9"             
#> [151] "AB(R)"            
#> 
#> $`hemibrain:v1.0`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                     
#> [1,] "GET"  "/api/v:ver/npexplorer/simpleconnections"
#> [2,] "POST" "/api/npexplorer/distribution"           
#> [3,] "GET"  "/oauth2callback"                        
#> [4,] "PUT"  "/api"                                   
#> [5,] "GET"  "/api/npexplorer/rankedtable"            
#> [6,] "POST" "/api/v:ver/npexplorer/autapses"         
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"
#> [3,] "github.com/janelia-flyem/echo-secure.oauthCallbackHandler"                                 
#> [4,] "github.com/labstack/echo.glob..func1"                                                      
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getRankedTable-fm" 
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }