Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $fib19
#> $fib19$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $fib19$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $fib19$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $fib19$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $fib19$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $fib19$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "NO3(L)"            "g5(R)"             "AL-DC3(R)"        
#>   [4] "AL-VL2a(R)"        "IB"                "AL-DM1(R)"        
#>   [7] "AL-VA1v(R)"        "AL-VA3(R)"         "AL-VM5v(R)"       
#>  [10] "MB(L)"             "FB-column3"        "LH(R)"            
#>  [13] "AL-DL3(R)"         "AL-VP4(R)"         "NO1(R)"           
#>  [16] "MB(+ACA)(R)"       "a3(R)"             "g2(R)"            
#>  [19] "AL-VA7m(R)"        "aL(L)"             "AL-VC2(R)"        
#>  [22] "AL-VA6(L)"         "PB(R1)"            "gL(R)"            
#>  [25] "PB(L7)"            "a'2(R)"            "AL-DM4(L)"        
#>  [28] "AMMC"              "PB(L1)"            "CA(L)"            
#>  [31] "a'3(R)"            "PB(L9)"            "IPS(R)"           
#>  [34] "INP"               "AL-DM2(R)"         "FBl1"             
#>  [37] "AL-VP1m(R)"        "AL-DP1l(R)"        "mALT(R)"          
#>  [40] "g4(R)"             "FBl4"              "AB(R)"            
#>  [43] "PENP"              "NO2(R)"            "AL-DL1(R)"        
#>  [46] "PB"                "AL-D(L)"           "AL-DL2v(R)"       
#>  [49] "EBr3d"             "AL-VM3(R)"         "RUB(R)"           
#>  [52] "vACA(R)"           "AL-DL5(L)"         "AL-DC1(R)"        
#>  [55] "AL-VP3(R)"         "AL-DL4(L)"         "a'1(R)"           
#>  [58] "CRE(-ROB,-RUB)(R)" "ATL(R)"            "bL(R)"            
#>  [61] "AL-VM7d(R)"        "b'2(R)"            "LX(R)"            
#>  [64] "gL(L)"             "FBl6"              "FLA(R)"           
#>  [67] "AL-DM6(L)"         "SLP(R)"            "g3(R)"            
#>  [70] "EBr5"              "ROB(R)"            "FBl7"             
#>  [73] "SMP(R)"            "AL-DM5(L)"         "AL-VA4(R)"        
#>  [76] "AL-VP1l(R)"        "PB(R5)"            "AOTU(R)"          
#>  [79] "CRE(L)"            "AL-DA2(L)"         "LAL(R)"           
#>  [82] "AL-DP1m(R)"        "AL-VM2(R)"         "SCL(L)"           
#>  [85] "PRW"               "lACA(R)"           "dACA(R)"          
#>  [88] "LO(R)"             "VES(R)"            "AL-VP5(R)"        
#>  [91] "PB(R8)"            "EPA(R)"            "NO"               
#>  [94] "IB112"             "CRE(-RUB)(L)"      "PB(L8)"           
#>  [97] "AL-DA3(R)"         "AL-DM3(R)"         "a2(R)"            
#> [100] "PED(R)"            "g1(R)"             "NO1(L)"           
#> [103] "AL-VM1(R)"         "BU(L)"             "PB(L6)"           
#> [106] "FBl3"              "PB(R2)"            "GC"               
#> [109] "SPS(L)"            "FBl2"              "AL-VM7v(R)"       
#> [112] "GA(R)"             "ME(R)"             "AL-DM2(L)"        
#> [115] "SAD(-AMMC)"        "b'L(R)"            "SNP(L)"           
#> [118] "AL-VA6(R)"         "AL-VL1(R)"         "mALT(L)"          
#> [121] "SMP(L)"            "AL-DM1(L)"         "AL-DL5(R)"        
#> [124] "AL-DM5(R)"         "AL-VM4(R)"         "AOT(R)"           
#> [127] "PB(L2)"            "PB(L3)"            "AL-DC2(L)"        
#> [130] "EBr3pw"            "AL-VC3l(R)"        "LX(L)"            
#> [133] "LOP(R)"            "AL(L)"             "GOR(R)"           
#> [136] "AL-DA3(L)"         "GNG"               "AL-VP2(R)"        
#> [139] "AB(L)"             "VMNP"              "AL-DP1m(L)"       
#> [142] "FBl9"              "AL-VC1(R)"         "WED(R)"           
#> [145] "SNP(R)"            "AL-VC5(R)"         "AL-DA4l(R)"       
#> [148] "AL-VL2p(R)"        "SCL(R)"            "PB(R4)"           
#> [151] "AL-DL4(R)"         "aL(R)"             "AL-VM7v(L)"       
#> [154] "b1(R)"             "AL(R)"             "AL-DC1(L)"        
#> [157] "AL-DM3(L)"         "AL-DM6(R)"         "VLNP(R)"          
#> [160] "EPA(L)"            "VES(L)"            "CX"               
#> [163] "FBl8"              "SAD"               "SIP(L)"           
#> [166] "PLP(R)"            "a'L(R)"            "LAL(L)"           
#> [169] "OL(R)"             "AL-D(R)"           "AL-VA7l(R)"       
#> [172] "bL(L)"             "a'L(L)"            "AL-DC2(R)"        
#> [175] "PB(R6)"            "EBr1"              "BU(R)"            
#> [178] "AL-VM7d(L)"        "SIP(R)"            "ICL(L)"           
#> [181] "AL-DA2(R)"         "CRE(R)"            "NO3(R)"           
#> [184] "EB"                "AL-DL2d(R)"        "a1(R)"            
#> [187] "GF(R)"             "b2(R)"             "AL-VA2(R)"        
#> [190] "AL-V(R)"           "PVLP(R)"           "ATL(L)"           
#> [193] "FBl5"              "AVLP(R)"           "NO(L)"            
#> [196] "ICL(R)"            "MB(R)"             "AL-DA4m(R)"       
#> [199] "CA(R)"             "SPS(R)"            "FB"               
#> [202] "NO2(L)"            "AME(R)"            "CAN(R)"           
#> [205] "AL-DA1(R)"         "PB(L5)"            "EBr6"             
#> [208] "AL-DM4(R)"         "EBr3am"            "EBr2r4"           
#> [211] "AL-DA4m(L)"        "b'L(L)"            "AL-VM5d(R)"       
#> [214] "AL-VP1d(R)"        "POC"               "PB(R7)"           
#> [217] "LAL(-GA)(R)"       "PB(R3)"            "RUB(L)"           
#> [220] "b'1(R)"            "PB(R9)"            "AL-VA1d(R)"       
#> [223] "AL-VC3m(R)"        "AL-VC4(R)"         "PB(L4)"           
#> [226] "AL-DC4(R)"         "AL-VA5(R)"         "hemibrain"        
#> [229] "GOR(L)"            "AL-DC4(L)"         "NO(R)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PRW"               "OL(R)"             "GOR(L)"           
#>   [4] "EBr5"              "PB(R9)"            "bL(L)"            
#>   [7] "a3(R)"             "b'L(L)"            "PB(L1)"           
#>  [10] "NO1(R)"            "CA(R)"             "IPS(R)"           
#>  [13] "CA(L)"             "a2(R)"             "GC"               
#>  [16] "VMNP"              "MB(+ACA)(R)"       "FBl9"             
#>  [19] "CRE(-ROB,-RUB)(R)" "a'2(R)"            "g4(R)"            
#>  [22] "SNP(L)"            "FBl6"              "SLP(R)"           
#>  [25] "PB(R6)"            "SMP(L)"            "AOTU(R)"          
#>  [28] "lACA(R)"           "SIP(R)"            "FBl4"             
#>  [31] "LAL(R)"            "PB(L3)"            "PED(R)"           
#>  [34] "PB(R8)"            "PB(L9)"            "b'2(R)"           
#>  [37] "SAD"               "FBl2"              "PB(L4)"           
#>  [40] "ATL(L)"            "AMMC"              "SCL(R)"           
#>  [43] "dACA(R)"           "NO"                "PB(L7)"           
#>  [46] "a'L(L)"            "IB"                "CX"               
#>  [49] "g3(R)"             "PB(R7)"            "NO(R)"            
#>  [52] "gL(R)"             "ICL(L)"            "FBl7"             
#>  [55] "INP"               "PENP"              "FB"               
#>  [58] "AVLP(R)"           "a'3(R)"            "GA(R)"            
#>  [61] "mALT(R)"           "SNP(R)"            "g5(R)"            
#>  [64] "MB(R)"             "SCL(L)"            "AOT(R)"           
#>  [67] "LAL(L)"            "CRE(R)"            "g2(R)"            
#>  [70] "ATL(R)"            "AB(R)"             "PB(R4)"           
#>  [73] "VES(L)"            "SPS(L)"            "PLP(R)"           
#>  [76] "ME(R)"             "ROB(R)"            "PB(L6)"           
#>  [79] "b1(R)"             "aL(R)"             "EBr2r4"           
#>  [82] "FLA(R)"            "EBr3pw"            "FBl5"             
#>  [85] "FBl8"              "SMP(R)"            "SIP(L)"           
#>  [88] "BU(L)"             "a1(R)"             "LX(R)"            
#>  [91] "FBl3"              "PB(R2)"            "EBr3d"            
#>  [94] "SAD(-AMMC)"        "AME(R)"            "RUB(L)"           
#>  [97] "bL(R)"             "AL(R)"             "PB(L8)"           
#> [100] "a'L(R)"            "CAN(R)"            "LH(R)"            
#> [103] "PB(R3)"            "EBr3am"            "vACA(R)"          
#> [106] "mALT(L)"           "b2(R)"             "EPA(R)"           
#> [109] "EPA(L)"            "PB(R5)"            "LAL(-GA)(R)"      
#> [112] "NO1(L)"            "AL(L)"             "CRE(L)"           
#> [115] "a'1(R)"            "PVLP(R)"           "gL(L)"            
#> [118] "POC"               "NO3(R)"            "MB(L)"            
#> [121] "PB(R1)"            "FBl1"              "NO2(R)"           
#> [124] "b'1(R)"            "NO2(L)"            "g1(R)"            
#> [127] "b'L(R)"            "EB"                "LX(L)"            
#> [130] "GNG"               "hemibrain"         "AB(L)"            
#> [133] "NO3(L)"            "aL(L)"             "PB(L2)"           
#> [136] "RUB(R)"            "GF(R)"             "SPS(R)"           
#> [139] "WED(R)"            "BU(R)"             "LO(R)"            
#> [142] "ICL(R)"            "NO(L)"             "PB(L5)"           
#> [145] "EBr1"              "PB"                "VLNP(R)"          
#> [148] "LOP(R)"            "EBr6"              "VES(R)"           
#> [151] "GOR(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-VA6(L)"         "FBl5"              "FBl6"             
#>   [4] "AL-DL3(R)"         "AL-VM3(R)"         "CA(L)"            
#>   [7] "PLP(R)"            "gL(R)"             "AL-DA3(R)"        
#>  [10] "AL-DL5(R)"         "AL-DM4(L)"         "AL-VA1d(R)"       
#>  [13] "NO2(L)"            "SNP(L)"            "CRE(R)"           
#>  [16] "FBl4"              "AL-DA1(R)"         "b'2(R)"           
#>  [19] "AVLP(R)"           "AL-DL2d(R)"        "AL-VA4(R)"        
#>  [22] "AL-VM1(R)"         "AL-VP1m(R)"        "a1(R)"            
#>  [25] "WED(R)"            "AL-DL4(L)"         "AL-DL5(L)"        
#>  [28] "LH(R)"             "AL-DC4(R)"         "PB(R6)"           
#>  [31] "AL-DM2(L)"         "PB(L4)"            "GA(R)"            
#>  [34] "EPA(R)"            "FBl3"              "AOTU(R)"          
#>  [37] "SCL(R)"            "EPA(L)"            "VLNP(R)"          
#>  [40] "PENP"              "NO2(R)"            "AL-DL2v(R)"       
#>  [43] "AL-VM7d(R)"        "AOT(R)"            "LAL(-GA)(R)"      
#>  [46] "GNG"               "PB(R5)"            "EB"               
#>  [49] "b2(R)"             "NO3(R)"            "FBl7"             
#>  [52] "AL-VC2(R)"         "AL-VL1(R)"         "SIP(L)"           
#>  [55] "AL-VP1l(R)"        "a'L(R)"            "AL(L)"            
#>  [58] "b1(R)"             "AL-VM4(R)"         "POC"              
#>  [61] "CX"                "GC"                "PB(R4)"           
#>  [64] "NO1(L)"            "AL-VC4(R)"         "AL-VM2(R)"        
#>  [67] "AL-DC4(L)"         "GOR(L)"            "CRE(L)"           
#>  [70] "FB-column3"        "IPS(R)"            "SAD"              
#>  [73] "LO(R)"             "AL-VP5(R)"         "RUB(R)"           
#>  [76] "PB"                "a'1(R)"            "AB(L)"            
#>  [79] "BU(R)"             "PB(R1)"            "VES(R)"           
#>  [82] "AL-VA3(R)"         "ROB(R)"            "PB(L8)"           
#>  [85] "PB(L3)"            "RUB(L)"            "ICL(R)"           
#>  [88] "AL-DC1(R)"         "PRW"               "a3(R)"            
#>  [91] "SCL(L)"            "PB(L1)"            "PB(L2)"           
#>  [94] "AL-VP2(R)"         "a2(R)"             "hemibrain"        
#>  [97] "VES(L)"            "AL-DP1l(R)"        "AL-VA7m(R)"       
#> [100] "AL-V(R)"           "SMP(L)"            "AL-VA7l(R)"       
#> [103] "g2(R)"             "AL-DM2(R)"         "AL-DM4(R)"        
#> [106] "AL-VM5d(R)"        "EBr3am"            "OL(R)"            
#> [109] "PB(L6)"            "FBl1"              "a'L(L)"           
#> [112] "LAL(L)"            "aL(L)"             "AL-VM7v(L)"       
#> [115] "gL(L)"             "AL-DA2(L)"         "AL-DM1(L)"        
#> [118] "AMMC"              "EBr3d"             "AL-D(L)"          
#> [121] "AL-D(R)"           "AL-VP4(R)"         "EBr2r4"           
#> [124] "GOR(R)"            "PB(R2)"            "MB(L)"            
#> [127] "EBr5"              "FBl2"              "PB(L5)"           
#> [130] "lACA(R)"           "b'L(R)"            "ICL(L)"           
#> [133] "AL-VA1v(R)"        "AL-VA2(R)"         "FB"               
#> [136] "b'L(L)"            "PED(R)"            "AL-DM5(L)"        
#> [139] "dACA(R)"           "AL-DA4m(L)"        "SIP(R)"           
#> [142] "NO3(L)"            "AB(R)"             "AL-DM3(R)"        
#> [145] "AL-VL2a(R)"        "SPS(R)"            "LOP(R)"           
#> [148] "PB(R3)"            "AL-VA6(R)"         "LX(R)"            
#> [151] "PB(L7)"            "AL-DL4(R)"         "g1(R)"            
#> [154] "VMNP"              "CRE(-RUB)(L)"      "EBr3pw"           
#> [157] "PVLP(R)"           "EBr6"              "a'2(R)"           
#> [160] "AL-DC2(L)"         "AL-DM3(L)"         "LAL(R)"           
#> [163] "vACA(R)"           "PB(R8)"            "g4(R)"            
#> [166] "PB(L9)"            "bL(L)"             "PB(R7)"           
#> [169] "NO(R)"             "FBl8"              "AL-VC3l(R)"       
#> [172] "AL-VP3(R)"         "SPS(L)"            "AL-DM6(L)"        
#> [175] "MB(+ACA)(R)"       "AL-DA2(R)"         "AL-DL1(R)"        
#> [178] "ME(R)"             "MB(R)"             "AL-DM5(R)"        
#> [181] "AL-VM5v(R)"        "mALT(R)"           "AL-DA4l(R)"       
#> [184] "a'3(R)"            "CRE(-ROB,-RUB)(R)" "NO"               
#> [187] "AL-DA4m(R)"        "AL-DC1(L)"         "CA(R)"            
#> [190] "NO(L)"             "g3(R)"             "AL-VC1(R)"        
#> [193] "AL-VP1d(R)"        "NO1(R)"            "AL-DA3(L)"        
#> [196] "AL(R)"             "EBr1"              "g5(R)"            
#> [199] "ATL(R)"            "SLP(R)"            "AL-VC5(R)"        
#> [202] "IB"                "BU(L)"             "AL-VM7v(R)"       
#> [205] "ATL(L)"            "AL-VA5(R)"         "FBl9"             
#> [208] "AL-DM1(R)"         "GF(R)"             "SMP(R)"           
#> [211] "AL-DC2(R)"         "PB(R9)"            "AL-DC3(R)"        
#> [214] "b'1(R)"            "SAD(-AMMC)"        "AME(R)"           
#> [217] "CAN(R)"            "AL-DM6(R)"         "AL-DP1m(R)"       
#> [220] "mALT(L)"           "AL-DP1m(L)"        "AL-VM7d(L)"       
#> [223] "LX(L)"             "INP"               "bL(R)"            
#> [226] "FLA(R)"            "SNP(R)"            "AL-VC3m(R)"       
#> [229] "AL-VL2p(R)"        "aL(R)"            
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "SPS(L)"            "FBl1"              "a'L(L)"           
#>   [4] "AL-VA7m(R)"        "AL-VM7d(R)"        "PB(R5)"           
#>   [7] "AB(L)"             "EB"                "AL-DM1(L)"        
#>  [10] "AL-VA7l(R)"        "FB"                "EPA(R)"           
#>  [13] "INP"               "GNG"               "PB(R7)"           
#>  [16] "AL-DA3(L)"         "bL(L)"             "FB-column3"       
#>  [19] "GOR(L)"            "OL(R)"             "AL-DP1l(R)"       
#>  [22] "AL-DM5(L)"         "SIP(R)"            "AL-DA4m(R)"       
#>  [25] "AL-DC3(R)"         "AL-VA2(R)"         "PB(L1)"           
#>  [28] "AL-DC1(L)"         "AL(R)"             "AL-V(R)"          
#>  [31] "GA(R)"             "hemibrain"         "b'L(R)"           
#>  [34] "ROB(R)"            "CA(R)"             "AL-DA2(R)"        
#>  [37] "AL-VM5d(R)"        "NO(R)"             "b'2(R)"           
#>  [40] "SIP(L)"            "EBr5"              "FLA(R)"           
#>  [43] "EBr3d"             "AL-DA2(L)"         "AL-DM3(L)"        
#>  [46] "a3(R)"             "AL-D(R)"           "AL-VA4(R)"        
#>  [49] "AL-VC5(R)"         "ATL(L)"            "g3(R)"            
#>  [52] "AL-DM4(L)"         "AL-DC1(R)"         "a'3(R)"           
#>  [55] "GC"                "FBl4"              "AL-VC3l(R)"       
#>  [58] "SNP(R)"            "g4(R)"             "WED(R)"           
#>  [61] "LX(R)"             "AL-VM7v(L)"        "AL-VA6(R)"        
#>  [64] "AL-VM5v(R)"        "AL-VM7v(R)"        "NO2(L)"           
#>  [67] "LAL(L)"            "gL(L)"             "CAN(R)"           
#>  [70] "AL-DL2v(R)"        "gL(R)"             "EPA(L)"           
#>  [73] "PB(R4)"            "POC"               "IB"               
#>  [76] "AL-DM5(R)"         "SMP(L)"            "AL-VC4(R)"        
#>  [79] "VLNP(R)"           "SAD"               "LAL(-GA)(R)"      
#>  [82] "AL-DP1m(L)"        "SLP(R)"            "AL-VP1m(R)"       
#>  [85] "MB(+ACA)(R)"       "FBl5"              "AL-DM1(R)"        
#>  [88] "a1(R)"             "GF(R)"             "AOTU(R)"          
#>  [91] "AME(R)"            "NO3(L)"            "LO(R)"            
#>  [94] "LH(R)"             "EBr3am"            "FBl6"             
#>  [97] "AL-DM3(R)"         "PB(L2)"            "AL-VA1d(R)"       
#> [100] "CRE(L)"            "AL-VL2a(R)"        "aL(R)"            
#> [103] "g2(R)"             "NO1(R)"            "PB(R9)"           
#> [106] "AL-VP1d(R)"        "PLP(R)"            "AL-DA4m(L)"       
#> [109] "ICL(R)"            "VES(L)"            "mALT(L)"          
#> [112] "g1(R)"             "AL-DA3(R)"         "AL-DL3(R)"        
#> [115] "CRE(R)"            "g5(R)"             "MB(L)"            
#> [118] "MB(R)"             "aL(L)"             "b'L(L)"           
#> [121] "AL-VA5(R)"         "AL-VP5(R)"         "a'L(R)"           
#> [124] "AL-DM2(L)"         "AL-VA6(L)"         "VES(R)"           
#> [127] "FBl9"              "PB(R3)"            "IPS(R)"           
#> [130] "AL-VP4(R)"         "SMP(R)"            "mALT(R)"          
#> [133] "PB(L4)"            "AL-DA4l(R)"        "PB(R1)"           
#> [136] "FBl7"              "GOR(R)"            "AL-VM7d(L)"       
#> [139] "ATL(R)"            "AL-DM6(R)"         "EBr6"             
#> [142] "SNP(L)"            "NO(L)"             "AL-DC4(R)"        
#> [145] "AL-VA1v(R)"        "PB(L5)"            "AVLP(R)"          
#> [148] "RUB(L)"            "LX(L)"             "LAL(R)"           
#> [151] "AL-DL4(R)"         "a'2(R)"            "PB(R2)"           
#> [154] "b2(R)"             "SCL(R)"            "FBl3"             
#> [157] "LOP(R)"            "PENP"              "NO1(L)"           
#> [160] "NO2(R)"            "a2(R)"             "EBr1"             
#> [163] "AL-DC2(L)"         "AL-D(L)"           "PB(L9)"           
#> [166] "PVLP(R)"           "AB(R)"             "PB(R8)"           
#> [169] "NO"                "EBr3pw"            "AL-DL5(R)"        
#> [172] "AL-DP1m(R)"        "AL-VL2p(R)"        "AOT(R)"           
#> [175] "AL-VM3(R)"         "BU(L)"             "VMNP"             
#> [178] "AMMC"              "AL-DL2d(R)"        "AL-VM2(R)"        
#> [181] "b'1(R)"            "SPS(R)"            "PB(L7)"           
#> [184] "FBl2"              "CRE(-ROB,-RUB)(R)" "CA(L)"            
#> [187] "AL-DL5(L)"         "PED(R)"            "AL-DC4(L)"        
#> [190] "AL-DC2(R)"         "AL-VC1(R)"         "AL-VM4(R)"        
#> [193] "PB(L3)"            "FBl8"              "AL-DL1(R)"        
#> [196] "SCL(L)"            "lACA(R)"           "AL(L)"            
#> [199] "AL-DL4(L)"         "bL(R)"             "AL-DM2(R)"        
#> [202] "AL-VC3m(R)"        "AL-VP3(R)"         "vACA(R)"          
#> [205] "PB(L8)"            "a'1(R)"            "ICL(L)"           
#> [208] "AL-VA3(R)"         "AL-VC2(R)"         "AL-VL1(R)"        
#> [211] "dACA(R)"           "SAD(-AMMC)"        "AL-DA1(R)"        
#> [214] "NO3(R)"            "PB"                "AL-DM6(L)"        
#> [217] "CRE(-RUB)(L)"      "AL-DM4(R)"         "ME(R)"            
#> [220] "BU(R)"             "PRW"               "AL-VP2(R)"        
#> [223] "AL-VP1l(R)"        "CX"                "PB(L6)"           
#> [226] "b1(R)"             "AL-VM1(R)"         "RUB(R)"           
#> [229] "PB(R6)"            "EBr2r4"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method  path                                        
#> [1,] "TRACE" "/api/*"                                    
#> [2,] "POST"  "/api/v:ver/npexplorer/neuronmeta"          
#> [3,] "GET"   "/api/npexplorer/roisinneuron"              
#> [4,] "POST"  "/api/v:ver/raw/cypher/transaction/:id/kill"
#> [5,] "GET"   "/api/v:ver/npexplorer/rankedtable"         
#> [6,] "GET"   "/api/npexplorer/simpleconnections"         
#>      name                                                                                             
#> [1,] "github.com/labstack/echo/v4.glob..func1"                                                        
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"     
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }