Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "VES(L)"            "PB(R5)"            "PB(L5)"           
#>   [4] "CX"                "SMP(L)"            "PB(L1)"           
#>   [7] "b'L(R)"            "ROB(R)"            "PENP"             
#>  [10] "a2(R)"             "GOR(R)"            "FB"               
#>  [13] "aL(L)"             "dACA(R)"           "AB(L)"            
#>  [16] "VLNP(R)"           "PB(R8)"            "OL(R)"            
#>  [19] "MB(+ACA)(R)"       "FBl5"              "NO(L)"            
#>  [22] "MB(R)"             "g3(R)"             "ICL(L)"           
#>  [25] "LO(R)"             "GF(R)"             "EBr3d"            
#>  [28] "SAD(-AMMC)"        "CAN(R)"            "LOP(R)"           
#>  [31] "GA(R)"             "EBr3pw"            "a'L(R)"           
#>  [34] "bL(L)"             "CA(R)"             "AME(R)"           
#>  [37] "LAL(L)"            "lACA(R)"           "EB"               
#>  [40] "SCL(L)"            "PB(R7)"            "aL(R)"            
#>  [43] "IPS(R)"            "EBr1"              "gL(L)"            
#>  [46] "VES(R)"            "FBl7"              "b'2(R)"           
#>  [49] "mALT(R)"           "mALT(L)"           "NO2(L)"           
#>  [52] "g2(R)"             "PB(R2)"            "ME(R)"            
#>  [55] "g4(R)"             "a'1(R)"            "FBl4"             
#>  [58] "ATL(R)"            "a3(R)"             "EBr6"             
#>  [61] "PB"                "a1(R)"             "RUB(L)"           
#>  [64] "CRE(R)"            "g5(R)"             "SIP(R)"           
#>  [67] "CRE(L)"            "FBl8"              "LH(R)"            
#>  [70] "AB(R)"             "LAL(-GA)(R)"       "EPA(R)"           
#>  [73] "NO3(R)"            "SCL(R)"            "PB(L8)"           
#>  [76] "SNP(R)"            "AL(L)"             "gL(R)"            
#>  [79] "PED(R)"            "GOR(L)"            "SNP(L)"           
#>  [82] "AL(R)"             "PB(L9)"            "ICL(R)"           
#>  [85] "SLP(R)"            "PLP(R)"            "NO"               
#>  [88] "NO3(L)"            "FBl1"              "FBl3"             
#>  [91] "FBl9"              "EBr2r4"            "NO(R)"            
#>  [94] "FBl6"              "a'L(L)"            "AOT(R)"           
#>  [97] "b'1(R)"            "CRE(-ROB,-RUB)(R)" "PB(R9)"           
#> [100] "FBl2"              "PB(R4)"            "b'L(L)"           
#> [103] "PB(L2)"            "a'3(R)"            "g1(R)"            
#> [106] "INP"               "NO1(L)"            "BU(R)"            
#> [109] "PB(R3)"            "LAL(R)"            "vACA(R)"          
#> [112] "AOTU(R)"           "SAD"               "GC"               
#> [115] "VMNP"              "PB(L7)"            "EBr3am"           
#> [118] "PB(L3)"            "SIP(L)"            "PB(L6)"           
#> [121] "LX(L)"             "IB"                "hemibrain"        
#> [124] "a'2(R)"            "FLA(R)"            "POC"              
#> [127] "ATL(L)"            "EPA(L)"            "SPS(R)"           
#> [130] "BU(L)"             "PB(R1)"            "AVLP(R)"          
#> [133] "SMP(R)"            "PRW"               "b2(R)"            
#> [136] "LX(R)"             "PVLP(R)"           "bL(R)"            
#> [139] "RUB(R)"            "CA(L)"             "PB(R6)"           
#> [142] "WED(R)"            "SPS(L)"            "NO2(R)"           
#> [145] "MB(L)"             "EBr5"              "GNG"              
#> [148] "PB(L4)"            "AMMC"              "NO1(R)"           
#> [151] "b1(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                           
#> [1,] "POST" "/api/npexplorer/findneurons"                  
#> [2,] "GET"  "/api/npexplorer/autapses"                     
#> [3,] "GET"  "/api/v:ver/npexplorer/celltype/:dataset/:type"
#> [4,] "POST" "/api/v:ver/raw/cypher/transaction/:id/kill"   
#> [5,] "POST" "/api/npexplorer/neuronmetavals"               
#> [6,] "POST" "/api/npexplorer/rankedtable"                  
#>      name                                                                                          
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getFindNeurons-fm"   
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"      
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                          
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMetaVals-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getRankedTable-fm"   

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }