Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "CX"                "lACA(R)"           "EBr1"             
#>   [4] "FLA(R)"            "mALT(L)"           "FBl1"             
#>   [7] "b'1(R)"            "a1(R)"             "GC"               
#>  [10] "CRE(R)"            "g2(R)"             "dACA(R)"          
#>  [13] "OL(R)"             "RUB(R)"            "LAL(L)"           
#>  [16] "SCL(R)"            "b'L(L)"            "NO(L)"            
#>  [19] "a'2(R)"            "AB(L)"             "a2(R)"            
#>  [22] "SPS(L)"            "a'L(R)"            "PVLP(R)"          
#>  [25] "CAN(R)"            "LAL(-GA)(R)"       "aL(R)"            
#>  [28] "PB(R8)"            "g4(R)"             "EBr3pw"           
#>  [31] "FBl9"              "CA(R)"             "POC"              
#>  [34] "vACA(R)"           "EB"                "NO2(R)"           
#>  [37] "bL(R)"             "AL(R)"             "LX(R)"            
#>  [40] "VES(L)"            "CRE(-ROB,-RUB)(R)" "VMNP"             
#>  [43] "ROB(R)"            "ICL(R)"            "LAL(R)"           
#>  [46] "EBr3am"            "g3(R)"             "PRW"              
#>  [49] "PB(R1)"            "NO(R)"             "SMP(R)"           
#>  [52] "LOP(R)"            "SAD(-AMMC)"        "LO(R)"            
#>  [55] "RUB(L)"            "GNG"               "FBl3"             
#>  [58] "a3(R)"             "AOT(R)"            "PB(L2)"           
#>  [61] "GOR(L)"            "g1(R)"             "IPS(R)"           
#>  [64] "LX(L)"             "PB(R2)"            "AME(R)"           
#>  [67] "VLNP(R)"           "IB"                "SPS(R)"           
#>  [70] "NO1(L)"            "NO1(R)"            "gL(R)"            
#>  [73] "MB(+ACA)(R)"       "b1(R)"             "FBl2"             
#>  [76] "FBl6"              "NO3(R)"            "b2(R)"            
#>  [79] "b'L(R)"            "SNP(R)"            "CA(L)"            
#>  [82] "SAD"               "PB(R7)"            "PB(L6)"           
#>  [85] "PB(R3)"            "AVLP(R)"           "EPA(L)"           
#>  [88] "MB(L)"             "ME(R)"             "SIP(R)"           
#>  [91] "FBl5"              "aL(L)"             "GOR(R)"           
#>  [94] "AMMC"              "FB"                "EBr5"             
#>  [97] "CRE(L)"            "PB(L7)"            "EPA(R)"           
#> [100] "g5(R)"             "AOTU(R)"           "PLP(R)"           
#> [103] "gL(L)"             "PB(L5)"            "GF(R)"            
#> [106] "PB(L8)"            "AB(R)"             "a'L(L)"           
#> [109] "b'2(R)"            "BU(L)"             "FBl7"             
#> [112] "SMP(L)"            "GA(R)"             "NO"               
#> [115] "FBl4"              "NO2(L)"            "PED(R)"           
#> [118] "PB(L1)"            "PB"                "ATL(L)"           
#> [121] "mALT(R)"           "EBr6"              "a'1(R)"           
#> [124] "PB(R5)"            "SLP(R)"            "MB(R)"            
#> [127] "PB(R6)"            "PB(L3)"            "EBr2r4"           
#> [130] "BU(R)"             "ATL(R)"            "PB(R4)"           
#> [133] "PENP"              "WED(R)"            "PB(R9)"           
#> [136] "LH(R)"             "FBl8"              "PB(L4)"           
#> [139] "SCL(L)"            "EBr3d"             "PB(L9)"           
#> [142] "bL(L)"             "ICL(L)"            "NO3(L)"           
#> [145] "VES(R)"            "INP"               "a'3(R)"           
#> [148] "hemibrain"         "SIP(L)"            "AL(L)"            
#> [151] "SNP(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                             
#> [1,] "GET"     "/login"                         
#> [2,] "POST"    "/api/npexplorer/roiconnectivity"
#> [3,] "TRACE"   "/api"                           
#> [4,] "GET"     "/api/npexplorer/distribution"   
#> [5,] "REPORT"  "/api"                           
#> [6,] "CONNECT" "/api/*"                         
#>      name                                                                                           
#> [1,] "github.com/janelia-flyem/echo-secure.loginHandler"                                            
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIConnectivity-fm"
#> [3,] "github.com/labstack/echo.glob..func1"                                                         
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"   
#> [5,] "github.com/labstack/echo.glob..func1"                                                         
#> [6,] "github.com/labstack/echo.glob..func1"                                                         

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }