Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0`
#> $`hemibrain:v1.0`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0`$ROIs
#>   [1] "PB(R3)"            "a3(R)"             "CA(L)"            
#>   [4] "LAL(-GA)(R)"       "g5(R)"             "MB(+ACA)(R)"      
#>   [7] "CA(R)"             "aL(L)"             "a'L(L)"           
#>  [10] "a2(R)"             "EB"                "EPA(R)"           
#>  [13] "PB(R1)"            "ICL(R)"            "SMP(R)"           
#>  [16] "INP"               "PENP"              "MB(L)"            
#>  [19] "g4(R)"             "EPA(L)"            "PB(L9)"           
#>  [22] "FBl3"              "LAL(L)"            "SAD"              
#>  [25] "b2(R)"             "PB(L4)"            "PB(L8)"           
#>  [28] "ICL(L)"            "FBl5"              "ATL(R)"           
#>  [31] "hemibrain"         "SPS(R)"            "b'L(R)"           
#>  [34] "EBr5"              "PB(R4)"            "VLNP(R)"          
#>  [37] "PB(R5)"            "FBl7"              "SCL(L)"           
#>  [40] "NO1(L)"            "a'L(R)"            "AL(L)"            
#>  [43] "LO(R)"             "a'3(R)"            "AB(L)"            
#>  [46] "a1(R)"             "EBr3pw"            "FBl9"             
#>  [49] "AVLP(R)"           "GNG"               "EBr3am"           
#>  [52] "AOTU(R)"           "FB"                "SAD(-AMMC)"       
#>  [55] "MB(R)"             "AL(R)"             "PB(L5)"           
#>  [58] "ATL(L)"            "SMP(L)"            "PED(R)"           
#>  [61] "LX(R)"             "NO3(L)"            "AB(R)"            
#>  [64] "RUB(R)"            "GC"                "PLP(R)"           
#>  [67] "NO"                "PB(L6)"            "IB"               
#>  [70] "NO(R)"             "a'1(R)"            "FBl8"             
#>  [73] "OL(R)"             "LH(R)"             "VES(L)"           
#>  [76] "SLP(R)"            "PRW"               "PB"               
#>  [79] "gL(L)"             "SNP(L)"            "PB(L3)"           
#>  [82] "GOR(R)"            "NO(L)"             "CX"               
#>  [85] "FBl2"              "bL(R)"             "LOP(R)"           
#>  [88] "GF(R)"             "WED(R)"            "CRE(L)"           
#>  [91] "PVLP(R)"           "b'1(R)"            "PB(L2)"           
#>  [94] "AMMC"              "SPS(L)"            "ME(R)"            
#>  [97] "SIP(R)"            "CAN(R)"            "PB(R7)"           
#> [100] "BU(R)"             "FBl4"              "b'L(L)"           
#> [103] "SCL(R)"            "BU(L)"             "EBr3d"            
#> [106] "PB(R9)"            "FBl6"              "AME(R)"           
#> [109] "RUB(L)"            "NO2(L)"            "NO3(R)"           
#> [112] "EBr1"              "ROB(R)"            "VES(R)"           
#> [115] "lACA(R)"           "mALT(R)"           "PB(L7)"           
#> [118] "FLA(R)"            "mALT(L)"           "EBr2r4"           
#> [121] "LAL(R)"            "g3(R)"             "g1(R)"            
#> [124] "CRE(R)"            "PB(L1)"            "PB(R8)"           
#> [127] "bL(L)"             "NO2(R)"            "CRE(-ROB,-RUB)(R)"
#> [130] "PB(R6)"            "aL(R)"             "b1(R)"            
#> [133] "vACA(R)"           "b'2(R)"            "g2(R)"            
#> [136] "a'2(R)"            "FBl1"              "gL(R)"            
#> [139] "LX(L)"             "VMNP"              "SNP(R)"           
#> [142] "GA(R)"             "GOR(L)"            "EBr6"             
#> [145] "NO1(R)"            "POC"               "AOT(R)"           
#> [148] "SIP(L)"            "dACA(R)"           "PB(R2)"           
#> [151] "IPS(R)"           
#> 
#> $`hemibrain:v1.0`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                          
#> [1,] "PUT"  "/api"                                        
#> [2,] "POST" "/api/npexplorer/distribution"                
#> [3,] "POST" "/api/npexplorer/completeness"                
#> [4,] "GET"  "/api/npexplorer/celltype/:dataset/:type"     
#> [5,] "HEAD" "/api/*"                                      
#> [6,] "POST" "/api/v:ver/raw/cypher/transaction/:id/cypher"
#>      name                                                                                        
#> [1,] "github.com/labstack/echo.glob..func1"                                                      
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCompleteness-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCellType-fm"    
#> [5,] "github.com/labstack/echo.glob..func1"                                                      
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }