Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AVLP(R)"           "PB(L9)"            "AOTU(R)"          
#>   [4] "MB(R)"             "AL(R)"             "AOT(R)"           
#>   [7] "PB(L3)"            "AL-VM2(R)"         "a1(R)"            
#>  [10] "AL-DA3(R)"         "AL-DL5(R)"         "AL-VP3(R)"        
#>  [13] "EB"                "NO(L)"             "LAL(R)"           
#>  [16] "AL-DL4(R)"         "PB(R5)"            "VMNP"             
#>  [19] "PB(L6)"            "a3(R)"             "AL-DM3(R)"        
#>  [22] "g3(R)"             "AL-DC1(R)"         "AL-VC2(R)"        
#>  [25] "AL-VM3(R)"         "AL-VM5d(R)"        "SIP(L)"           
#>  [28] "GF(R)"             "VES(R)"            "FBl9"             
#>  [31] "AL-VP1m(R)"        "LAL(L)"            "BU(R)"            
#>  [34] "g1(R)"             "AL-VA5(R)"         "AL-VA6(R)"        
#>  [37] "RUB(R)"            "gL(R)"             "AL-DA4l(R)"       
#>  [40] "NO2(L)"            "NO3(R)"            "AL-DP1m(L)"       
#>  [43] "AL-VM7d(L)"        "PB(R9)"            "PB(L7)"           
#>  [46] "EBr1"              "CAN(R)"            "AL-DM2(L)"        
#>  [49] "AL-DA2(R)"         "SMP(L)"            "AL-VM7v(L)"       
#>  [52] "NO3(L)"            "PRW"               "GNG"              
#>  [55] "AL-VM7d(R)"        "AL-DA3(L)"         "AL-DL3(R)"        
#>  [58] "FBl6"              "EBr3d"             "FBl1"             
#>  [61] "PB(R3)"            "OL(R)"             "AL-VP1l(R)"       
#>  [64] "AB(L)"             "EBr3am"            "dACA(R)"          
#>  [67] "AL-VM4(R)"         "b'1(R)"            "LOP(R)"           
#>  [70] "b'L(R)"            "PLP(R)"            "AL-DL1(R)"        
#>  [73] "aL(R)"             "SAD(-AMMC)"        "SMP(R)"           
#>  [76] "CX"                "AL-D(L)"           "PED(R)"           
#>  [79] "AL-DC4(L)"         "ATL(R)"            "POC"              
#>  [82] "b'2(R)"            "SNP(R)"            "AL-VC3l(R)"       
#>  [85] "AL-VM5v(R)"        "BU(L)"             "ROB(R)"           
#>  [88] "NO1(R)"            "ICL(L)"            "g2(R)"            
#>  [91] "NO(R)"             "b'L(L)"            "aL(L)"            
#>  [94] "AL-DP1m(R)"        "mALT(L)"           "AL-DM3(L)"        
#>  [97] "AL-VC3m(R)"        "AL-VP1d(R)"        "a2(R)"            
#> [100] "SCL(R)"            "PB"                "a'2(R)"           
#> [103] "AL-DM4(L)"         "PB(R7)"            "FBl3"             
#> [106] "FBl7"              "EPA(R)"            "EBr6"             
#> [109] "EBr2r4"            "EBr3pw"            "PVLP(R)"          
#> [112] "PB(R4)"            "AL-DM5(R)"         "FB"               
#> [115] "AL(L)"             "CRE(L)"            "FBl2"             
#> [118] "AL-VA4(R)"         "NO1(L)"            "AL-VA7l(R)"       
#> [121] "CA(L)"             "AL-DM1(L)"         "AL-DM5(L)"        
#> [124] "CRE(R)"            "RUB(L)"            "a'L(L)"           
#> [127] "AL-DP1l(R)"        "PB(L2)"            "SAD"              
#> [130] "ICL(R)"            "AL-DC3(R)"         "AL-VM7v(R)"       
#> [133] "PB(R8)"            "AME(R)"            "AL-DM6(L)"        
#> [136] "b1(R)"             "AB(R)"             "hemibrain"        
#> [139] "a'L(R)"            "AL-DM4(R)"         "LX(R)"            
#> [142] "FLA(R)"            "SNP(L)"            "IB112"            
#> [145] "vACA(R)"           "WED(R)"            "PB(L1)"           
#> [148] "bL(L)"             "AL-VC4(R)"         "AL-VC5(R)"        
#> [151] "AL-VL2p(R)"        "PENP"              "AL-DC4(R)"        
#> [154] "LO(R)"             "AL-VP5(R)"         "AL-DL5(L)"        
#> [157] "AL-DC2(R)"         "AL-DM2(R)"         "MB(+ACA)(R)"      
#> [160] "FBl5"              "AL-DL2d(R)"        "AL-DL2v(R)"       
#> [163] "PB(R1)"            "MB(L)"             "AL-VA6(L)"        
#> [166] "FBl8"              "PB(R2)"            "LX(L)"            
#> [169] "AL-VC1(R)"         "AL-VP4(R)"         "GC"               
#> [172] "SCL(L)"            "AL-VA1d(R)"        "CRE(-RUB)(L)"     
#> [175] "AL-DA2(L)"         "AL-DA4m(L)"        "EPA(L)"           
#> [178] "SPS(R)"            "SLP(R)"            "AL-VA1v(R)"       
#> [181] "SIP(R)"            "AL-DL4(L)"         "NO"               
#> [184] "LH(R)"             "AL-D(R)"           "AL-VL1(R)"        
#> [187] "a'3(R)"            "CRE(-ROB,-RUB)(R)" "gL(L)"            
#> [190] "VES(L)"            "AL-DM6(R)"         "PB(R6)"           
#> [193] "SPS(L)"            "FB-column3"        "IPS(R)"           
#> [196] "ATL(L)"            "AMMC"              "lACA(R)"          
#> [199] "LAL(-GA)(R)"       "EBr5"              "g4(R)"            
#> [202] "bL(R)"             "AL-VA3(R)"         "PB(L5)"           
#> [205] "GA(R)"             "AL-DC1(L)"         "AL-DC2(L)"        
#> [208] "INP"               "AL-VL2a(R)"        "b2(R)"            
#> [211] "AL-VP2(R)"         "VLNP(R)"           "mALT(R)"          
#> [214] "AL-DM1(R)"         "AL-VA2(R)"         "g5(R)"            
#> [217] "CA(R)"             "NO2(R)"            "AL-V(R)"          
#> [220] "FBl4"              "AL-DA1(R)"         "AL-VM1(R)"        
#> [223] "IB"                "GOR(R)"            "PB(L8)"           
#> [226] "ME(R)"             "a'1(R)"            "PB(L4)"           
#> [229] "AL-DA4m(R)"        "AL-VA7m(R)"        "GOR(L)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "VMNP"              "PB(R9)"            "PB(L9)"           
#>   [4] "FBl5"              "PB(R7)"            "dACA(R)"          
#>   [7] "FBl4"              "AL(R)"             "LOP(R)"           
#>  [10] "AMMC"              "LAL(-GA)(R)"       "aL(L)"            
#>  [13] "GOR(R)"            "AOT(R)"            "CX"               
#>  [16] "b1(R)"             "LAL(L)"            "aL(R)"            
#>  [19] "b2(R)"             "LO(R)"             "FBl7"             
#>  [22] "PENP"              "a1(R)"             "PRW"              
#>  [25] "GC"                "BU(L)"             "ROB(R)"           
#>  [28] "PED(R)"            "g1(R)"             "EBr5"             
#>  [31] "SPS(L)"            "FBl9"              "AOTU(R)"          
#>  [34] "hemibrain"         "MB(L)"             "SNP(R)"           
#>  [37] "EPA(R)"            "MB(+ACA)(R)"       "ATL(R)"           
#>  [40] "INP"               "PB(R6)"            "MB(R)"            
#>  [43] "VES(L)"            "SCL(L)"            "lACA(R)"          
#>  [46] "PB(R1)"            "CRE(L)"            "EBr3am"           
#>  [49] "WED(R)"            "ICL(L)"            "b'L(L)"           
#>  [52] "VLNP(R)"           "a'3(R)"            "SPS(R)"           
#>  [55] "NO1(R)"            "PB(L6)"            "EBr3pw"           
#>  [58] "NO3(R)"            "NO(R)"             "SNP(L)"           
#>  [61] "CA(R)"             "AVLP(R)"           "PB(L4)"           
#>  [64] "AB(L)"             "NO2(L)"            "PB(R8)"           
#>  [67] "PB(L3)"            "PLP(R)"            "SIP(R)"           
#>  [70] "GNG"               "CA(L)"             "SMP(L)"           
#>  [73] "gL(L)"             "RUB(R)"            "PB"               
#>  [76] "a'1(R)"            "POC"               "AL(L)"            
#>  [79] "GOR(L)"            "a'L(L)"            "IB"               
#>  [82] "GA(R)"             "ICL(R)"            "CAN(R)"           
#>  [85] "SLP(R)"            "VES(R)"            "FB"               
#>  [88] "PVLP(R)"           "PB(R4)"            "b'1(R)"           
#>  [91] "a2(R)"             "g2(R)"             "FLA(R)"           
#>  [94] "a'2(R)"            "FBl1"              "FBl2"             
#>  [97] "LH(R)"             "GF(R)"             "gL(R)"            
#> [100] "FBl8"              "LAL(R)"            "OL(R)"            
#> [103] "EBr3d"             "b'L(R)"            "EBr1"             
#> [106] "SAD(-AMMC)"        "AME(R)"            "SMP(R)"           
#> [109] "CRE(R)"            "LX(L)"             "vACA(R)"          
#> [112] "FBl3"              "NO1(L)"            "EBr2r4"           
#> [115] "PB(L2)"            "LX(R)"             "SIP(L)"           
#> [118] "g5(R)"             "FBl6"              "CRE(-ROB,-RUB)(R)"
#> [121] "ATL(L)"            "b'2(R)"            "g4(R)"            
#> [124] "bL(L)"             "PB(R2)"            "RUB(L)"           
#> [127] "PB(R5)"            "EBr6"              "PB(L1)"           
#> [130] "PB(L8)"            "ME(R)"             "PB(R3)"           
#> [133] "AB(R)"             "NO2(R)"            "g3(R)"            
#> [136] "PB(L7)"            "NO(L)"             "mALT(L)"          
#> [139] "mALT(R)"           "SCL(R)"            "NO3(L)"           
#> [142] "bL(R)"             "a3(R)"             "SAD"              
#> [145] "a'L(R)"            "BU(R)"             "NO"               
#> [148] "IPS(R)"            "EPA(L)"            "PB(L5)"           
#> [151] "EB"               
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "g2(R)"             "PB(R2)"            "LAL(-GA)(R)"      
#>   [4] "PB(R8)"            "FBl2"              "AL-DM6(R)"        
#>   [7] "AL-VA7l(R)"        "CRE(R)"            "EBr5"             
#>  [10] "FBl5"              "AL-VM2(R)"         "b'L(R)"           
#>  [13] "b1(R)"             "SLP(R)"            "LOP(R)"           
#>  [16] "g1(R)"             "NO1(R)"            "NO2(R)"           
#>  [19] "AL-DM4(R)"         "AL-VC1(R)"         "b'L(L)"           
#>  [22] "PB(R5)"            "NO"                "CRE(L)"           
#>  [25] "AL-VM5v(R)"        "PENP"              "PB(L6)"           
#>  [28] "AL-VA4(R)"         "NO2(L)"            "a'L(R)"           
#>  [31] "g4(R)"             "AL-D(L)"           "AL(R)"            
#>  [34] "AL-DL1(R)"         "PVLP(R)"           "a'3(R)"           
#>  [37] "NO(L)"             "VLNP(R)"           "ATL(L)"           
#>  [40] "hemibrain"         "NO1(L)"            "PB(L8)"           
#>  [43] "AL-DA3(L)"         "AB(R)"             "SPS(L)"           
#>  [46] "AME(R)"            "AL-VC2(R)"         "CX"               
#>  [49] "EBr3am"            "PB(L5)"            "GC"               
#>  [52] "NO3(R)"            "ME(R)"             "g5(R)"            
#>  [55] "FBl1"              "AL-DM2(L)"         "bL(R)"            
#>  [58] "EB"                "CAN(R)"            "SCL(L)"           
#>  [61] "PB(R1)"            "aL(L)"             "PB(R6)"           
#>  [64] "IPS(R)"            "PB(R4)"            "AL-DA3(R)"        
#>  [67] "AL-DM3(R)"         "FLA(R)"            "AL-VM1(R)"        
#>  [70] "LX(R)"             "EBr2r4"            "AL-DM6(L)"        
#>  [73] "VES(L)"            "AL-DA2(R)"         "GNG"              
#>  [76] "AL-VM4(R)"         "b'2(R)"            "SMP(L)"           
#>  [79] "LO(R)"             "ICL(L)"            "RUB(L)"           
#>  [82] "LAL(L)"            "CRE(-RUB)(L)"      "VES(R)"           
#>  [85] "AL-DA4m(R)"        "AL-VL2a(R)"        "gL(L)"            
#>  [88] "SIP(R)"            "ICL(R)"            "AL-DL4(R)"        
#>  [91] "AL-VA1d(R)"        "SCL(R)"            "PB(L1)"           
#>  [94] "LX(L)"             "MB(L)"             "VMNP"             
#>  [97] "PB(R9)"            "PB(L4)"            "AL-VC3m(R)"       
#> [100] "vACA(R)"           "FBl4"              "AL-DM5(R)"        
#> [103] "AL-VA3(R)"         "AL-VM7d(R)"        "AL-VP1m(R)"       
#> [106] "dACA(R)"           "AL-DC4(L)"         "EBr3pw"           
#> [109] "FBl7"              "mALT(L)"           "AL-V(R)"          
#> [112] "EPA(R)"            "AL-DA4m(L)"        "CA(R)"            
#> [115] "ATL(R)"            "LAL(R)"            "PLP(R)"           
#> [118] "AL-DL3(R)"         "AL-VA1v(R)"        "PB(L2)"           
#> [121] "AL-DC1(L)"         "a'1(R)"            "FBl3"             
#> [124] "AL-VC5(R)"         "AL-VP1d(R)"        "PED(R)"           
#> [127] "AL-DC2(L)"         "PB(R3)"            "AL-VA2(R)"        
#> [130] "EBr6"              "SNP(R)"            "FB-column3"       
#> [133] "AL-DL5(R)"         "CRE(-ROB,-RUB)(R)" "FB"               
#> [136] "SNP(L)"            "OL(R)"             "AL-DC4(R)"        
#> [139] "AL-VA5(R)"         "AL-DL2d(R)"        "SMP(R)"           
#> [142] "AMMC"              "AL-VA6(L)"         "FBl8"             
#> [145] "AL-VA6(R)"         "AL-VM7v(R)"        "POC"              
#> [148] "lACA(R)"           "AL-DM3(L)"         "bL(L)"            
#> [151] "AL-DA1(R)"         "g3(R)"             "AL-DC2(R)"        
#> [154] "AL-DL5(L)"         "INP"               "a'L(L)"           
#> [157] "AL-VC3l(R)"        "AL(L)"             "MB(R)"            
#> [160] "AL-DA4l(R)"        "AL-DP1l(R)"        "AL-DP1m(R)"       
#> [163] "b'1(R)"            "AL-DM1(R)"         "AOT(R)"           
#> [166] "a2(R)"             "EBr3d"             "NO(R)"            
#> [169] "AB(L)"             "WED(R)"            "PB(L9)"           
#> [172] "GOR(R)"            "AL-VM3(R)"         "GA(R)"            
#> [175] "PRW"               "BU(L)"             "PB"               
#> [178] "NO3(L)"            "AOTU(R)"           "AL-DM1(L)"        
#> [181] "AL-VL2p(R)"        "AL-DP1m(L)"        "b2(R)"            
#> [184] "SAD(-AMMC)"        "AVLP(R)"           "GOR(L)"           
#> [187] "AL-DM4(L)"         "FBl9"              "AL-VP2(R)"        
#> [190] "SAD"               "PB(R7)"            "AL-DM5(L)"        
#> [193] "AL-DC3(R)"         "AL-VL1(R)"         "mALT(R)"          
#> [196] "LH(R)"             "AL-VP4(R)"         "a'2(R)"           
#> [199] "MB(+ACA)(R)"       "EPA(L)"            "aL(R)"            
#> [202] "PB(L7)"            "a3(R)"             "AL-VC4(R)"        
#> [205] "GF(R)"             "AL-D(R)"           "RUB(R)"           
#> [208] "a1(R)"             "SIP(L)"            "EBr1"             
#> [211] "ROB(R)"            "gL(R)"             "AL-VP1l(R)"       
#> [214] "AL-DA2(L)"         "AL-DL4(L)"         "AL-VM5d(R)"       
#> [217] "AL-DC1(R)"         "AL-DL2v(R)"        "AL-DM2(R)"        
#> [220] "PB(L3)"            "SPS(R)"            "AL-VM7d(L)"       
#> [223] "AL-VM7v(L)"        "FBl6"              "AL-VP3(R)"        
#> [226] "IB"                "CA(L)"             "BU(R)"            
#> [229] "AL-VA7m(R)"        "AL-VP5(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "PRW"               "bL(L)"             "SIP(L)"           
#>   [4] "AL-DM4(R)"         "AL-VA7l(R)"        "POC"              
#>   [7] "PB(L5)"            "NO1(R)"            "AL-DM4(L)"        
#>  [10] "PB(L4)"            "AL-VM7d(R)"        "AL-DL2d(R)"       
#>  [13] "PED(R)"            "aL(R)"             "a'1(R)"           
#>  [16] "FBl8"              "AL-DA2(R)"         "AL-DM3(R)"        
#>  [19] "AL-DA4m(L)"        "ME(R)"             "a'L(R)"           
#>  [22] "BU(R)"             "AL-VM7v(L)"        "LO(R)"            
#>  [25] "AL-DL5(L)"         "ROB(R)"            "AL-VP1l(R)"       
#>  [28] "vACA(R)"           "PB(L6)"            "AL-DL4(R)"        
#>  [31] "g2(R)"             "RUB(R)"            "a1(R)"            
#>  [34] "g5(R)"             "bL(R)"             "AL-VM7v(R)"       
#>  [37] "AL(L)"             "AL-DC1(L)"         "AL-VM7d(L)"       
#>  [40] "GNG"               "OL(R)"             "EBr2r4"           
#>  [43] "MB(+ACA)(R)"       "AL-VP5(R)"         "SNP(L)"           
#>  [46] "EBr3am"            "CX"                "AL-D(L)"          
#>  [49] "VES(R)"            "CA(R)"             "AL-VA3(R)"        
#>  [52] "a2(R)"             "FLA(R)"            "PB(R3)"           
#>  [55] "MB(R)"             "FBl3"              "AL-VC3m(R)"       
#>  [58] "AB(L)"             "CAN(R)"            "VES(L)"           
#>  [61] "AL-DC3(R)"         "AL-DM6(R)"         "PB(L8)"           
#>  [64] "AL-DM3(L)"         "NO2(R)"            "AL-DL1(R)"        
#>  [67] "AOTU(R)"           "gL(R)"             "AL-DC1(R)"        
#>  [70] "AL-V(R)"           "PB(R8)"            "AL-DA3(R)"        
#>  [73] "PB(L1)"            "PB(R9)"            "PVLP(R)"          
#>  [76] "ATL(R)"            "b'L(L)"            "NO"               
#>  [79] "LAL(-GA)(R)"       "AL-VA1v(R)"        "GF(R)"            
#>  [82] "PB(L3)"            "NO1(L)"            "a'2(R)"           
#>  [85] "RUB(L)"            "g1(R)"             "LX(R)"            
#>  [88] "EB"                "LAL(L)"            "FBl5"             
#>  [91] "PB(R2)"            "LOP(R)"            "LX(L)"            
#>  [94] "b'2(R)"            "FBl1"              "AL-DM2(R)"        
#>  [97] "AL-VA4(R)"         "IB"                "b'L(R)"           
#> [100] "EBr6"              "AL-DM1(L)"         "ICL(R)"           
#> [103] "AB(R)"             "AL-VA1d(R)"        "SCL(L)"           
#> [106] "AL-DC2(L)"         "EBr3pw"            "AL-DM1(R)"        
#> [109] "SIP(R)"            "ICL(L)"            "FBl7"             
#> [112] "hemibrain"         "SAD"               "AL(R)"            
#> [115] "AL-VA6(R)"         "AL-VC2(R)"         "NO(R)"            
#> [118] "AVLP(R)"           "AL-DL5(R)"         "CRE(L)"           
#> [121] "IPS(R)"            "AL-VL2p(R)"        "AL-VM2(R)"        
#> [124] "AL-VP3(R)"         "a'L(L)"            "AL-DC2(R)"        
#> [127] "AL-VM5d(R)"        "AL-DC4(R)"         "mALT(L)"          
#> [130] "PB(R6)"            "SPS(R)"            "AL-DA4m(R)"       
#> [133] "AL-VL2a(R)"        "lACA(R)"           "CRE(R)"           
#> [136] "VMNP"              "INP"               "AL-DP1l(R)"       
#> [139] "EBr3d"             "AL-DP1m(L)"        "b'1(R)"           
#> [142] "SMP(L)"            "mALT(R)"           "g4(R)"            
#> [145] "GOR(R)"            "AL-DM2(L)"         "EPA(L)"           
#> [148] "AL-VA5(R)"         "AL-VC1(R)"         "g3(R)"            
#> [151] "b2(R)"             "MB(L)"             "BU(L)"            
#> [154] "AL-DL4(L)"         "SAD(-AMMC)"        "AL-VM4(R)"        
#> [157] "GC"                "FB"                "LH(R)"            
#> [160] "AOT(R)"            "CA(L)"             "gL(L)"            
#> [163] "SLP(R)"            "AL-VC3l(R)"        "EBr1"             
#> [166] "EBr5"              "SNP(R)"            "AL-DA3(L)"        
#> [169] "AL-DL3(R)"         "AL-VM3(R)"         "VLNP(R)"          
#> [172] "GOR(L)"            "PB(L7)"            "b1(R)"            
#> [175] "AL-VA2(R)"         "GA(R)"             "AL-VA6(L)"        
#> [178] "aL(L)"             "AL-VC4(R)"         "AL-VM5v(R)"       
#> [181] "NO2(L)"            "PB"                "AL-VP1m(R)"       
#> [184] "PB(R7)"            "dACA(R)"           "PLP(R)"           
#> [187] "CRE(-RUB)(L)"      "FBl6"              "a3(R)"            
#> [190] "AL-VP2(R)"         "SCL(R)"            "PB(R4)"           
#> [193] "AL-DA1(R)"         "PENP"              "AL-DM6(L)"        
#> [196] "NO3(L)"            "FBl4"              "AL-DM5(R)"        
#> [199] "AL-VM1(R)"         "SMP(R)"            "a'3(R)"           
#> [202] "ATL(L)"            "AMMC"              "SPS(L)"           
#> [205] "FBl2"              "AL-DA4l(R)"        "NO3(R)"           
#> [208] "AL-DC4(L)"         "PB(L9)"            "FB-column3"       
#> [211] "CRE(-ROB,-RUB)(R)" "EPA(R)"            "PB(R1)"           
#> [214] "AL-DL2v(R)"        "PB(R5)"            "AL-DA2(L)"        
#> [217] "AME(R)"            "AL-DP1m(R)"        "AL-D(R)"          
#> [220] "AL-VL1(R)"         "FBl9"              "NO(L)"            
#> [223] "LAL(R)"            "AL-VA7m(R)"        "AL-VP1d(R)"       
#> [226] "PB(L2)"            "WED(R)"            "AL-DM5(L)"        
#> [229] "AL-VC5(R)"         "AL-VP4(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method     path                                
#> [1,] "POST"     "/api/v:ver/npexplorer/rankedtable" 
#> [2,] "PROPFIND" "/api/*"                            
#> [3,] "GET"      "/*"                                
#> [4,] "GET"      "/api/v:ver/cached/roiconnectivity" 
#> [5,] "PATCH"    "/api"                              
#> [6,] "POST"     "/api/v:ver/npexplorer/roisinneuron"
#>      name                                                                
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [2,] "github.com/labstack/echo.glob..func1"                              
#> [3,] "github.com/labstack/echo.common.static.func1"                      
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [5,] "github.com/labstack/echo.glob..func1"                              
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }