Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0`
#> $`hemibrain:v1.0`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0`$ROIs
#>   [1] "a1(R)"             "b'L(R)"            "PB(R9)"           
#>   [4] "VES(R)"            "INP"               "PVLP(R)"          
#>   [7] "PB(R5)"            "RUB(R)"            "EBr3am"           
#>  [10] "b'2(R)"            "LAL(-GA)(R)"       "EBr5"             
#>  [13] "gL(L)"             "dACA(R)"           "LO(R)"            
#>  [16] "bL(R)"             "VLNP(R)"           "a'1(R)"           
#>  [19] "SCL(R)"            "g4(R)"             "EPA(R)"           
#>  [22] "AL(L)"             "PB(L8)"            "PB(L9)"           
#>  [25] "LOP(R)"            "a'3(R)"            "PB(R6)"           
#>  [28] "FB"                "ICL(R)"            "b1(R)"            
#>  [31] "PLP(R)"            "a'2(R)"            "SLP(R)"           
#>  [34] "PB(R4)"            "b'1(R)"            "lACA(R)"          
#>  [37] "GOR(L)"            "PB(R2)"            "PB(R3)"           
#>  [40] "FBl8"              "PRW"               "SPS(R)"           
#>  [43] "NO3(R)"            "EBr3pw"            "FLA(R)"           
#>  [46] "ATL(R)"            "AB(L)"             "PENP"             
#>  [49] "SIP(L)"            "PB(L1)"            "ATL(L)"           
#>  [52] "ROB(R)"            "AMMC"              "OL(R)"            
#>  [55] "LAL(L)"            "PB(R7)"            "EBr1"             
#>  [58] "a3(R)"             "BU(L)"             "AME(R)"           
#>  [61] "CAN(R)"            "SCL(L)"            "GA(R)"            
#>  [64] "WED(R)"            "IB"                "CRE(-ROB,-RUB)(R)"
#>  [67] "FBl4"              "FBl7"              "FBl9"             
#>  [70] "PB(L4)"            "mALT(R)"           "EBr3d"            
#>  [73] "gL(R)"             "g5(R)"             "SAD"              
#>  [76] "PB(R8)"            "NO"                "LX(L)"            
#>  [79] "hemibrain"         "aL(R)"             "PB(L7)"           
#>  [82] "CA(L)"             "IPS(R)"            "RUB(L)"           
#>  [85] "EBr2r4"            "FBl1"              "AVLP(R)"          
#>  [88] "AB(R)"             "FBl2"              "FBl5"             
#>  [91] "g2(R)"             "PB"                "ME(R)"            
#>  [94] "SAD(-AMMC)"        "MB(+ACA)(R)"       "FBl6"             
#>  [97] "LH(R)"             "VES(L)"            "MB(L)"            
#> [100] "EBr6"              "LX(R)"             "NO3(L)"           
#> [103] "MB(R)"             "AL(R)"             "CX"               
#> [106] "g1(R)"             "PB(R1)"            "bL(L)"            
#> [109] "ICL(L)"            "AOT(R)"            "PB(L2)"           
#> [112] "AOTU(R)"           "a2(R)"             "SPS(L)"           
#> [115] "SIP(R)"            "NO1(L)"            "a'L(R)"           
#> [118] "BU(R)"             "SMP(R)"            "SNP(R)"           
#> [121] "SNP(L)"            "vACA(R)"           "NO2(L)"           
#> [124] "CRE(R)"            "b2(R)"             "GOR(R)"           
#> [127] "PB(L6)"            "CRE(L)"            "NO2(R)"           
#> [130] "NO(L)"             "g3(R)"             "PB(L3)"           
#> [133] "VMNP"              "EB"                "LAL(R)"           
#> [136] "NO(R)"             "FBl3"              "CA(R)"            
#> [139] "b'L(L)"            "GNG"               "a'L(L)"           
#> [142] "GF(R)"             "SMP(L)"            "GC"               
#> [145] "NO1(R)"            "PED(R)"            "aL(L)"            
#> [148] "EPA(L)"            "mALT(L)"           "POC"              
#> [151] "PB(L5)"           
#> 
#> $`hemibrain:v1.0`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                     
#> [1,] "GET"  "/api/dbmeta/database"                   
#> [2,] "POST" "/api/*"                                 
#> [3,] "GET"  "/api/cached/roiconnectivity"            
#> [4,] "POST" "/api/v:ver/roimeshes/mesh/:dataset/:roi"
#> [5,] "GET"  "/api/v:ver/npexplorer/findneurons"      
#> [6,] "GET"  "/api/serverinfo"                        
#>      name                                                                                    
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatabase-fm"        
#> [2,] "github.com/labstack/echo.glob..func1"                                                  
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROIConnectivity-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [6,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }