Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "FBl6"              "EBr6"              "SNP(R)"           
#>   [4] "AL-VL1(R)"         "PB(R6)"            "GF(R)"            
#>   [7] "PB(R7)"            "AL-DM6(L)"         "FBl2"             
#>  [10] "PB(R2)"            "FBl5"              "AL-VM1(R)"        
#>  [13] "g2(R)"             "EBr1"              "PB(L8)"           
#>  [16] "AL-VC4(R)"         "PB(L5)"            "CRE(-ROB,-RUB)(R)"
#>  [19] "ICL(L)"            "AL-D(R)"           "AL-VA4(R)"        
#>  [22] "PB(L3)"            "gL(L)"             "AL-DC2(L)"        
#>  [25] "VES(R)"            "AVLP(R)"           "PB(R5)"           
#>  [28] "AOTU(R)"           "MB(L)"             "a3(R)"            
#>  [31] "AL-DL5(R)"         "AL-VA1v(R)"        "EBr3pw"           
#>  [34] "GOR(R)"            "AL-DP1l(R)"        "AL-VP4(R)"        
#>  [37] "PLP(R)"            "AL-DA4m(L)"        "NO(R)"            
#>  [40] "AL-D(L)"           "FB-column3"        "INP"              
#>  [43] "AL-DM5(R)"         "IPS(R)"            "AL-VM7v(R)"       
#>  [46] "PB(R1)"            "aL(L)"             "AL-DM3(R)"        
#>  [49] "LOP(R)"            "EPA(R)"            "SLP(R)"           
#>  [52] "LO(R)"             "LX(L)"             "GNG"              
#>  [55] "b'L(L)"            "RUB(R)"            "dACA(R)"          
#>  [58] "AL-VA6(L)"         "MB(+ACA)(R)"       "FBl3"             
#>  [61] "OL(R)"             "vACA(R)"           "AL-DA3(L)"        
#>  [64] "RUB(L)"            "AL-VP1m(R)"        "AL-VP1d(R)"       
#>  [67] "VLNP(R)"           "AL-VM7d(L)"        "AB(R)"            
#>  [70] "AL-DL1(R)"         "SIP(R)"            "NO"               
#>  [73] "SAD(-AMMC)"        "AL-VP3(R)"         "hemibrain"        
#>  [76] "SIP(L)"            "AL-DA2(L)"         "AL-DM5(L)"        
#>  [79] "AL-DL2v(R)"        "SPS(L)"            "PB(R4)"           
#>  [82] "AL-DA4m(R)"        "b'L(R)"            "SCL(R)"           
#>  [85] "CA(R)"             "NO2(R)"            "VES(L)"           
#>  [88] "AL-DP1m(R)"        "AL-VA7m(R)"        "PB(L2)"           
#>  [91] "GA(R)"             "CRE(-RUB)(L)"      "ICL(R)"           
#>  [94] "AL-DL3(R)"         "AL-VP2(R)"         "WED(R)"           
#>  [97] "NO3(L)"            "GC"                "NO3(R)"           
#> [100] "CRE(L)"            "IB112"             "AME(R)"           
#> [103] "AOT(R)"            "lACA(R)"           "PVLP(R)"          
#> [106] "LH(R)"             "AL-VA2(R)"         "IB"               
#> [109] "ATL(L)"            "a1(R)"             "a2(R)"            
#> [112] "CRE(R)"            "a'L(R)"            "PB(L6)"           
#> [115] "PB(L9)"            "AL-DC2(R)"         "SMP(L)"           
#> [118] "AL(L)"             "FBl4"              "bL(R)"            
#> [121] "AL-VA3(R)"         "SMP(R)"            "ROB(R)"           
#> [124] "AL-DA3(R)"         "AL-DC4(R)"         "EBr2r4"           
#> [127] "b1(R)"             "AL-V(R)"           "PED(R)"           
#> [130] "a'L(L)"            "AL-VP1l(R)"        "SAD"              
#> [133] "MB(R)"             "AL-DM2(R)"         "AL-VM5d(R)"       
#> [136] "SPS(R)"            "b'2(R)"            "AL-DC4(L)"        
#> [139] "FBl8"              "g3(R)"             "PRW"              
#> [142] "VMNP"              "AL-DL5(L)"         "AL-DC1(R)"        
#> [145] "g1(R)"             "AL-VM4(R)"         "b'1(R)"           
#> [148] "LAL(-GA)(R)"       "PB(R8)"            "PB"               
#> [151] "LAL(L)"            "gL(R)"             "AL-DL2d(R)"       
#> [154] "AL-VA7l(R)"        "aL(R)"             "AL-DA2(R)"        
#> [157] "AL-VM7d(R)"        "FBl1"              "a'3(R)"           
#> [160] "NO2(L)"            "PENP"              "AL-DA4l(R)"       
#> [163] "AL-VM5v(R)"        "g4(R)"             "CAN(R)"           
#> [166] "AL-DC3(R)"         "FLA(R)"            "a'2(R)"           
#> [169] "AL-VP5(R)"         "EBr5"              "PB(R9)"           
#> [172] "FBl9"              "ME(R)"             "AB(L)"            
#> [175] "AL-DC1(L)"         "AL-DL4(L)"         "AL-DM1(L)"        
#> [178] "AL-DM4(R)"         "AL-VC3l(R)"        "EBr3am"           
#> [181] "AL-DM1(R)"         "AL-VA1d(R)"        "AL-VC1(R)"        
#> [184] "AL-VM3(R)"         "mALT(R)"           "b2(R)"            
#> [187] "PB(L7)"            "PB(R3)"            "LX(R)"            
#> [190] "EBr3d"             "ATL(R)"            "POC"              
#> [193] "FB"                "AL-DP1m(L)"        "a'1(R)"           
#> [196] "AL-DM4(L)"         "SNP(L)"            "AL(R)"            
#> [199] "SCL(L)"            "BU(L)"             "AL-VL2p(R)"       
#> [202] "NO1(L)"            "AL-DM2(L)"         "AL-DM3(L)"        
#> [205] "AL-VM7v(L)"        "bL(L)"             "AL-VL2a(R)"       
#> [208] "PB(L4)"            "AL-DA1(R)"         "AMMC"             
#> [211] "GOR(L)"            "EPA(L)"            "LAL(R)"           
#> [214] "AL-DL4(R)"         "BU(R)"             "g5(R)"            
#> [217] "AL-DM6(R)"         "AL-VA5(R)"         "AL-VA6(R)"        
#> [220] "AL-VC3m(R)"        "AL-VM2(R)"         "PB(L1)"           
#> [223] "CX"                "NO1(R)"            "NO(L)"            
#> [226] "AL-VC5(R)"         "mALT(L)"           "EB"               
#> [229] "FBl7"              "AL-VC2(R)"         "CA(L)"            
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "LAL(L)"            "EBr1"              "IPS(R)"           
#>   [4] "b'1(R)"            "CX"                "ROB(R)"           
#>   [7] "SAD(-AMMC)"        "AME(R)"            "a3(R)"            
#>  [10] "PB(R7)"            "FB"                "SNP(R)"           
#>  [13] "a'1(R)"            "FBl2"              "PB(L7)"           
#>  [16] "INP"               "PRW"               "lACA(R)"          
#>  [19] "PB(R8)"            "ME(R)"             "PB(L6)"           
#>  [22] "a'3(R)"            "b'L(R)"            "LX(R)"            
#>  [25] "a'L(R)"            "EPA(R)"            "PB(R1)"           
#>  [28] "SPS(R)"            "NO3(R)"            "PB(L2)"           
#>  [31] "PB(L3)"            "mALT(R)"           "FBl5"             
#>  [34] "FBl7"              "ATL(R)"            "SMP(L)"           
#>  [37] "a2(R)"             "RUB(L)"            "PB(R4)"           
#>  [40] "AMMC"              "SCL(R)"            "PB"               
#>  [43] "PLP(R)"            "gL(L)"             "EBr3pw"           
#>  [46] "gL(R)"             "FBl1"              "FBl6"             
#>  [49] "GOR(R)"            "PENP"              "PB(R6)"           
#>  [52] "LAL(-GA)(R)"       "NO1(L)"            "CRE(L)"           
#>  [55] "ATL(L)"            "AL(R)"             "SMP(R)"           
#>  [58] "g3(R)"             "VES(R)"            "ICL(R)"           
#>  [61] "b1(R)"             "bL(R)"             "GF(R)"            
#>  [64] "SIP(L)"            "PVLP(R)"           "FBl9"             
#>  [67] "PB(L4)"            "EBr3d"             "aL(L)"            
#>  [70] "AB(L)"             "vACA(R)"           "EPA(L)"           
#>  [73] "GNG"               "NO(L)"             "BU(L)"            
#>  [76] "NO1(R)"            "BU(R)"             "SIP(R)"           
#>  [79] "ICL(L)"            "OL(R)"             "SLP(R)"           
#>  [82] "a1(R)"             "VMNP"              "FBl8"             
#>  [85] "IB"                "PB(L5)"            "aL(R)"            
#>  [88] "EB"                "MB(+ACA)(R)"       "PB(R3)"           
#>  [91] "AB(R)"             "AOTU(R)"           "GA(R)"            
#>  [94] "dACA(R)"           "AVLP(R)"           "LAL(R)"           
#>  [97] "WED(R)"            "EBr2r4"            "VLNP(R)"          
#> [100] "g2(R)"             "a'2(R)"            "SNP(L)"           
#> [103] "FBl4"              "CA(R)"             "PB(L1)"           
#> [106] "FLA(R)"            "CA(L)"             "SPS(L)"           
#> [109] "PB(R2)"            "POC"               "mALT(L)"          
#> [112] "bL(L)"             "LO(R)"             "MB(R)"            
#> [115] "PED(R)"            "GOR(L)"            "CRE(R)"           
#> [118] "MB(L)"             "g5(R)"             "AOT(R)"           
#> [121] "b'2(R)"            "hemibrain"         "NO3(L)"           
#> [124] "CAN(R)"            "a'L(L)"            "SCL(L)"           
#> [127] "CRE(-ROB,-RUB)(R)" "g1(R)"             "NO"               
#> [130] "NO(R)"             "LX(L)"             "b'L(L)"           
#> [133] "NO2(L)"            "GC"                "AL(L)"            
#> [136] "LH(R)"             "VES(L)"            "PB(R5)"           
#> [139] "PB(L8)"            "PB(R9)"            "FBl3"             
#> [142] "NO2(R)"            "LOP(R)"            "g4(R)"            
#> [145] "PB(L9)"            "RUB(R)"            "EBr3am"           
#> [148] "SAD"               "b2(R)"             "EBr6"             
#> [151] "EBr5"             
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DL1(R)"         "VES(R)"            "b1(R)"            
#>   [4] "AL-DC2(R)"         "AL-DM1(L)"         "OL(R)"            
#>   [7] "FBl9"              "AL-DM6(R)"         "VMNP"             
#>  [10] "SCL(L)"            "LOP(R)"            "PB(L1)"           
#>  [13] "FB-column3"        "AL-VM7d(R)"        "AL-D(L)"          
#>  [16] "AL-DA4l(R)"        "AL-VA7l(R)"        "AL-DP1l(R)"       
#>  [19] "lACA(R)"           "SLP(R)"            "AL-DM3(R)"        
#>  [22] "AL(R)"             "AL-VA5(R)"         "AL-VA7m(R)"       
#>  [25] "AL-DC4(L)"         "aL(L)"             "INP"              
#>  [28] "g4(R)"             "AL-VA1v(R)"        "AL-V(R)"          
#>  [31] "a1(R)"             "FLA(R)"            "PLP(R)"           
#>  [34] "SNP(R)"            "AL-VP3(R)"         "g3(R)"            
#>  [37] "g2(R)"             "VES(L)"            "FBl3"             
#>  [40] "IPS(R)"            "AL-VC5(R)"         "PENP"             
#>  [43] "b'L(R)"            "NO"                "LAL(-GA)(R)"      
#>  [46] "AME(R)"            "AL-DP1m(R)"        "AL-DM3(L)"        
#>  [49] "AL-DP1m(L)"        "PED(R)"            "EBr6"             
#>  [52] "SPS(L)"            "AL-DL4(R)"         "AL-DL5(R)"        
#>  [55] "bL(R)"             "hemibrain"         "BU(R)"            
#>  [58] "AL(L)"             "SIP(R)"            "AL-VC3l(R)"       
#>  [61] "CA(L)"             "NO3(R)"            "bL(L)"            
#>  [64] "AB(L)"             "AL-VM5v(R)"        "AL-DL4(L)"        
#>  [67] "AL-VM7d(L)"        "FBl4"              "ATL(L)"           
#>  [70] "dACA(R)"           "RUB(L)"            "AL-DM1(R)"        
#>  [73] "AL-VC4(R)"         "LO(R)"             "PB(R3)"           
#>  [76] "AL-DC3(R)"         "PB(R2)"            "PB(R7)"           
#>  [79] "g1(R)"             "CRE(-RUB)(L)"      "PB(R9)"           
#>  [82] "CA(R)"             "a2(R)"             "NO2(L)"           
#>  [85] "AL-VM7v(L)"        "AL-DL2v(R)"        "AL-VP1m(R)"       
#>  [88] "PB(L2)"            "GC"                "GNG"              
#>  [91] "AL-DA4m(R)"        "AL-VA2(R)"         "AL-VP4(R)"        
#>  [94] "GOR(L)"            "WED(R)"            "a'L(L)"           
#>  [97] "EBr2r4"            "AL-VP5(R)"         "LAL(L)"           
#> [100] "AL-VA1d(R)"        "FB"                "FBl1"             
#> [103] "SMP(L)"            "PB(L6)"            "POC"              
#> [106] "AL-VA6(L)"         "AL-VC1(R)"         "AL-VM7v(R)"       
#> [109] "ICL(L)"            "PB(L7)"            "EBr1"             
#> [112] "LX(R)"             "AL-DM2(L)"         "RUB(R)"           
#> [115] "EBr3am"            "aL(R)"             "AL-VA4(R)"        
#> [118] "AL-VL2a(R)"        "LAL(R)"            "AL-VC3m(R)"       
#> [121] "CRE(R)"            "AVLP(R)"           "AL-DL2d(R)"       
#> [124] "AL-VM5d(R)"        "AL-VP2(R)"         "CX"               
#> [127] "NO1(R)"            "AL-DM5(R)"         "AL-VC2(R)"        
#> [130] "EBr3pw"            "AL-DC1(R)"         "GF(R)"            
#> [133] "SPS(R)"            "EBr5"              "NO(R)"            
#> [136] "SNP(L)"            "AL-VL1(R)"         "CRE(-ROB,-RUB)(R)"
#> [139] "AMMC"              "AOTU(R)"           "SIP(L)"           
#> [142] "AL-DM6(L)"         "GA(R)"             "NO1(L)"           
#> [145] "a'1(R)"            "FBl8"              "mALT(R)"          
#> [148] "AL-VM4(R)"         "IB"                "PB"               
#> [151] "EPA(R)"            "NO(L)"             "AL-VA6(R)"        
#> [154] "SCL(R)"            "AL-DL3(R)"         "AL-VA3(R)"        
#> [157] "a'3(R)"            "FBl5"              "FBl6"             
#> [160] "AL-VM2(R)"         "MB(L)"             "AL-D(R)"          
#> [163] "PB(L5)"            "AL-DA2(L)"         "GOR(R)"           
#> [166] "AL-DM2(R)"         "CRE(L)"            "AL-VM1(R)"        
#> [169] "AL-VP1l(R)"        "vACA(R)"           "ROB(R)"           
#> [172] "SAD(-AMMC)"        "AL-DA1(R)"         "b'1(R)"           
#> [175] "PB(R6)"            "EB"                "AOT(R)"           
#> [178] "gL(L)"             "PB(R1)"            "AB(R)"            
#> [181] "EPA(L)"            "PB(R4)"            "b'2(R)"           
#> [184] "PB(L3)"            "PB(L8)"            "NO2(R)"           
#> [187] "SAD"               "MB(+ACA)(R)"       "FBl2"             
#> [190] "LX(L)"             "AL-DA3(R)"         "AL-VL2p(R)"       
#> [193] "ME(R)"             "AL-DA3(L)"         "NO3(L)"           
#> [196] "AL-DA4m(L)"        "PVLP(R)"           "PB(L4)"           
#> [199] "PRW"               "AL-DM4(L)"         "CAN(R)"           
#> [202] "g5(R)"             "LH(R)"             "b'L(L)"           
#> [205] "MB(R)"             "SMP(R)"            "PB(R8)"           
#> [208] "AL-DL5(L)"         "PB(L9)"            "BU(L)"            
#> [211] "FBl7"              "AL-DM4(R)"         "AL-VM3(R)"        
#> [214] "EBr3d"             "gL(R)"             "ATL(R)"           
#> [217] "AL-VP1d(R)"        "PB(R5)"            "mALT(L)"          
#> [220] "a3(R)"             "AL-DC4(R)"         "VLNP(R)"          
#> [223] "b2(R)"             "a'L(R)"            "AL-DC2(L)"        
#> [226] "ICL(R)"            "AL-DA2(R)"         "a'2(R)"           
#> [229] "AL-DC1(L)"         "AL-DM5(L)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-DA2(L)"         "b'1(R)"            "AL(L)"            
#>   [4] "FBl2"              "SMP(R)"            "FB"               
#>   [7] "SAD(-AMMC)"        "AL-DM2(R)"         "SCL(L)"           
#>  [10] "ICL(L)"            "AME(R)"            "GOR(R)"           
#>  [13] "AL-VC4(R)"         "AOTU(R)"           "b'L(R)"           
#>  [16] "EPA(R)"            "AL-VP1m(R)"        "FBl6"             
#>  [19] "AL-DL4(R)"         "AL-VM5d(R)"        "POC"              
#>  [22] "a1(R)"             "FBl5"              "AL-VA7m(R)"       
#>  [25] "SMP(L)"            "GC"                "CRE(R)"           
#>  [28] "AL-DM3(R)"         "AL-DM5(R)"         "AL-VA3(R)"        
#>  [31] "SPS(L)"            "AL-VM7v(L)"        "FBl4"             
#>  [34] "PB(R4)"            "AL-VP5(R)"         "g3(R)"            
#>  [37] "MB(L)"             "AL-DM4(R)"         "AL-DM3(L)"        
#>  [40] "AL(R)"             "AL-DL2v(R)"        "LO(R)"            
#>  [43] "IPS(R)"            "FBl7"              "PENP"             
#>  [46] "LAL(-GA)(R)"       "AL-DM4(L)"         "AL-VP1l(R)"       
#>  [49] "SAD"               "EBr2r4"            "AL-D(R)"          
#>  [52] "AOT(R)"            "FLA(R)"            "AL-DL5(R)"        
#>  [55] "AL-VP3(R)"         "FBl3"              "AL-DA3(R)"        
#>  [58] "IB"                "mALT(R)"           "PB(L9)"           
#>  [61] "AL-VA1v(R)"        "AL-VC2(R)"         "GA(R)"            
#>  [64] "PRW"               "AL-DA2(R)"         "LH(R)"            
#>  [67] "AL-DL1(R)"         "AL-DL2d(R)"        "AL-VC5(R)"        
#>  [70] "AL-VL1(R)"         "VMNP"              "ME(R)"            
#>  [73] "PVLP(R)"           "AL-VM7v(R)"        "AL-VM7d(R)"       
#>  [76] "SIP(R)"            "AB(R)"             "NO(L)"            
#>  [79] "PB(R1)"            "a'1(R)"            "aL(L)"            
#>  [82] "aL(R)"             "gL(L)"             "AL-DM5(L)"        
#>  [85] "VES(R)"            "AMMC"              "SCL(R)"           
#>  [88] "AL-DM1(L)"         "AL-DM6(L)"         "g2(R)"            
#>  [91] "AL-DL4(L)"         "AL-DC2(L)"         "AL-VA6(L)"        
#>  [94] "bL(L)"             "b'L(L)"            "AL-VP2(R)"        
#>  [97] "NO3(R)"            "dACA(R)"           "PB(L1)"           
#> [100] "AL-VL2a(R)"        "ATL(R)"            "CX"               
#> [103] "NO"                "ROB(R)"            "FBl8"             
#> [106] "a'L(L)"            "AL-DA4m(R)"        "PLP(R)"           
#> [109] "gL(R)"             "AVLP(R)"           "EBr3d"            
#> [112] "AL-DM2(L)"         "AL-VM7d(L)"        "NO3(L)"           
#> [115] "NO2(R)"            "b'2(R)"            "PB(R7)"           
#> [118] "NO1(R)"            "AL-DC3(R)"         "mALT(L)"          
#> [121] "SNP(R)"            "PB(R9)"            "PB(L4)"           
#> [124] "LX(L)"             "AL-VC1(R)"         "LAL(L)"           
#> [127] "PB(L6)"            "AL-DA4l(R)"        "SLP(R)"           
#> [130] "AL-VP4(R)"         "SPS(R)"            "PB(L2)"           
#> [133] "VES(L)"            "AL-VC3m(R)"        "EBr3am"           
#> [136] "b2(R)"             "EBr5"              "RUB(R)"           
#> [139] "EBr1"              "NO(R)"             "EPA(L)"           
#> [142] "AL-DP1m(L)"        "PB(R2)"            "AL-VL2p(R)"       
#> [145] "AL-D(L)"           "AL-VA7l(R)"        "ATL(L)"           
#> [148] "LAL(R)"            "MB(R)"             "AL-VA5(R)"        
#> [151] "AL-DC4(R)"         "a2(R)"             "hemibrain"        
#> [154] "PED(R)"            "lACA(R)"           "LX(R)"            
#> [157] "EB"                "CRE(L)"            "INP"              
#> [160] "AL-DC1(R)"         "AL-DP1m(R)"        "AL-VM2(R)"        
#> [163] "PB(R6)"            "PB(L8)"            "FBl9"             
#> [166] "GF(R)"             "CAN(R)"            "AL-VA2(R)"        
#> [169] "b1(R)"             "AL-VA4(R)"         "PB(R5)"           
#> [172] "g1(R)"             "CRE(-RUB)(L)"      "g5(R)"            
#> [175] "OL(R)"             "CRE(-ROB,-RUB)(R)" "AL-DL3(R)"        
#> [178] "PB(L5)"            "CA(L)"             "a'2(R)"           
#> [181] "AL-DC4(L)"         "MB(+ACA)(R)"       "GOR(L)"           
#> [184] "AL-DC1(L)"         "AL-VM1(R)"         "AL-V(R)"          
#> [187] "a'3(R)"            "g4(R)"             "FBl1"             
#> [190] "bL(R)"             "PB"                "AL-DA4m(L)"       
#> [193] "AL-DM6(R)"         "AL-VA6(R)"         "vACA(R)"          
#> [196] "PB(L3)"            "WED(R)"            "PB(R8)"           
#> [199] "SNP(L)"            "AL-VM5v(R)"        "AL-VP1d(R)"       
#> [202] "AL-DA1(R)"         "AL-DP1l(R)"        "AL-VC3l(R)"       
#> [205] "EBr3pw"            "FB-column3"        "PB(R3)"           
#> [208] "AL-DL5(L)"         "a3(R)"             "AL-VM4(R)"        
#> [211] "EBr6"              "a'L(R)"            "AL-DA3(L)"        
#> [214] "AB(L)"             "GNG"               "AL-VA1d(R)"       
#> [217] "VLNP(R)"           "AL-DC2(R)"         "AL-DM1(R)"        
#> [220] "BU(L)"             "PB(L7)"            "RUB(L)"           
#> [223] "CA(R)"             "ICL(R)"            "AL-VM3(R)"        
#> [226] "LOP(R)"            "NO2(L)"            "SIP(L)"           
#> [229] "NO1(L)"            "BU(R)"            
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`manc:v1.0`
#> $`manc:v1.0`$`last-mod`
#> [1] "2023-05-31 23:54:01"
#> 
#> $`manc:v1.0`$uuid
#> [1] "59b37970bc7a4341b9a3a965a0d6b402"
#> 
#> $`manc:v1.0`$ROIs
#>  [1] "LegNp(T1)(L)"     "HTct(UTct-T3)(L)" "GF(L)"            "GF(R)"           
#>  [5] "ProLN(L)"         "LegNp(T2)(L)"     "WTct(UTct-T2)(L)" "DProN(L)"        
#>  [9] "mVAC(T3)(L)"      "WTct(UTct-T2)(R)" "ADMN(R)"          "VProN(L)"        
#> [13] "mVAC(T1)(R)"      "CvN(R)"           "HTct(UTct-T3)(R)" "LTct"            
#> [17] "AbN1(L)"          "AbN3(L)"          "AbNT"             "MesoAN(R)"       
#> [21] "CV"               "ADMN(L)"          "AbN2(R)"          "Ov(R)"           
#> [25] "DMetaN(L)"        "MetaLN(R)"        "ProAN(R)"         "Ov(L)"           
#> [29] "MesoLN(L)"        "mVAC(T2)(L)"      "mVAC(T2)(R)"      "IntTct"          
#> [33] "AbN1(R)"          "PDMN(L)"          "ProAN(L)"         "PrN(L)"          
#> [37] "PrN(R)"           "NTct(UTct-T1)(L)" "NTct(UTct-T1)(R)" "DProN(R)"        
#> [41] "ProLN(R)"         "DMetaN(R)"        "ProCN(L)"         "ProCN(R)"        
#> [45] "LegNp(T1)(R)"     "LegNp(T3)(L)"     "LegNp(T3)(R)"     "mVAC(T1)(L)"     
#> [49] "ANm"              "MesoLN(R)"        "mVAC(T3)(R)"      "AbN3(R)"         
#> [53] "AbN4(L)"          "AbN4(R)"          "VProN(R)"         "LegNp(T2)(R)"    
#> [57] "AbN2(L)"          "CvN(L)"           "MesoAN(L)"        "MetaLN(L)"       
#> [61] "PDMN(R)"         
#> 
#> $`manc:v1.0`$superLevelROIs
#>  [1] "ADMN(L)"          "ADMN(R)"          "Ov(L)"            "Ov(R)"           
#>  [5] "ANm"              "AbN1(L)"          "AbN1(R)"          "AbN2(L)"         
#>  [9] "AbN2(R)"          "AbN3(L)"          "AbN3(R)"          "AbN4(L)"         
#> [13] "AbN4(R)"          "AbNT"             "CV"               "CvN(L)"          
#> [17] "CvN(R)"           "DMetaN(L)"        "DMetaN(R)"        "DProN(L)"        
#> [21] "DProN(R)"         "GF(L)"            "GF(R)"            "HTct(UTct-T3)(L)"
#> [25] "HTct(UTct-T3)(R)" "LegNp(T1)(L)"     "LegNp(T1)(R)"     "LegNp(T2)(L)"    
#> [29] "LegNp(T2)(R)"     "LegNp(T3)(L)"     "LegNp(T3)(R)"     "IntTct"          
#> [33] "LTct"             "MesoAN(L)"        "MesoAN(R)"        "MesoLN(L)"       
#> [37] "MesoLN(R)"        "MetaLN(L)"        "MetaLN(R)"        "NTct(UTct-T1)(L)"
#> [41] "NTct(UTct-T1)(R)" "PDMN(L)"          "PrN(L)"           "PrN(R)"          
#> [45] "ProCN(L)"         "ProCN(R)"         "PDMN(R)"          "ProAN(L)"        
#> [49] "ProAN(R)"         "ProLN(L)"         "ProLN(R)"         "VProN(L)"        
#> [53] "VProN(R)"         "WTct(UTct-T2)(L)" "WTct(UTct-T2)(R)" "mVAC(T1)(L)"     
#> [57] "mVAC(T1)(R)"      "mVAC(T2)(L)"      "mVAC(T2)(R)"      "mVAC(T3)(L)"     
#> [61] "mVAC(T3)(R)"     
#> 
#> $`manc:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/manc-connectome"
#> 
#> $`manc:v1.0`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                     
#> [1,] "GET"  "/api/v:ver/dbmeta/instances"            
#> [2,] "POST" "/api/npexplorer/roisinneuron"           
#> [3,] "GET"  "/api/v:ver/roimeshes/mesh/:dataset/:roi"
#> [4,] "POST" "/api/raw/keyvalue/key/:instance/:key"   
#> [5,] "POST" "/api/v:ver/roimeshes/mesh/:dataset/:roi"
#> [6,] "POST" "/api/v:ver/npexplorer/neuronmeta"       
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }