Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "g1(R)"             "b'L(R)"            "dACA(R)"          
#>   [4] "gL(L)"             "ICL(R)"            "GOR(L)"           
#>   [7] "NO3(R)"            "CRE(R)"            "EBr5"             
#>  [10] "FLA(R)"            "PB(R8)"            "EBr3d"            
#>  [13] "CRE(L)"            "AOTU(R)"           "AMMC"             
#>  [16] "ICL(L)"            "SMP(L)"            "FBl6"             
#>  [19] "PB(L2)"            "SNP(L)"            "VMNP"             
#>  [22] "NO"                "PB(R5)"            "SCL(R)"           
#>  [25] "EBr1"              "EBr3pw"            "LAL(-GA)(R)"      
#>  [28] "PB(L6)"            "PB(L9)"            "VLNP(R)"          
#>  [31] "vACA(R)"           "MB(L)"             "SPS(R)"           
#>  [34] "gL(R)"             "MB(+ACA)(R)"       "OL(R)"            
#>  [37] "b'1(R)"            "a2(R)"             "a'L(R)"           
#>  [40] "BU(R)"             "PB(R9)"            "EBr3am"           
#>  [43] "PB(R6)"            "EPA(R)"            "NO3(L)"           
#>  [46] "LH(R)"             "CRE(-ROB,-RUB)(R)" "ATL(L)"           
#>  [49] "CX"                "PENP"              "mALT(R)"          
#>  [52] "EBr6"              "INP"               "IB"               
#>  [55] "hemibrain"         "PLP(R)"            "EBr2r4"           
#>  [58] "LAL(R)"            "a'L(L)"            "POC"              
#>  [61] "PRW"               "SIP(L)"            "AL(L)"            
#>  [64] "FBl1"              "FBl3"              "CAN(R)"           
#>  [67] "GOR(R)"            "PB(L5)"            "LAL(L)"           
#>  [70] "mALT(L)"           "PED(R)"            "GC"               
#>  [73] "g4(R)"             "ROB(R)"            "FBl4"             
#>  [76] "AB(R)"             "CA(L)"             "PB(R7)"           
#>  [79] "FBl8"              "LX(L)"             "NO(L)"            
#>  [82] "SMP(R)"            "FBl2"              "PB(L1)"           
#>  [85] "aL(R)"             "g3(R)"             "NO1(R)"           
#>  [88] "LX(R)"             "PVLP(R)"           "FBl5"             
#>  [91] "b'2(R)"            "VES(L)"            "PB(R1)"           
#>  [94] "bL(R)"             "AME(R)"            "RUB(R)"           
#>  [97] "NO2(R)"            "AL(R)"             "LOP(R)"           
#> [100] "FBl7"              "IPS(R)"            "a3(R)"            
#> [103] "SIP(R)"            "FBl9"              "PB(R2)"           
#> [106] "AVLP(R)"           "RUB(L)"            "SCL(L)"           
#> [109] "a'2(R)"            "NO1(L)"            "ME(R)"            
#> [112] "EB"                "NO(R)"             "GA(R)"            
#> [115] "g2(R)"             "VES(R)"            "PB(R4)"           
#> [118] "GNG"               "a'3(R)"            "PB(L3)"           
#> [121] "SAD"               "b2(R)"             "SNP(R)"           
#> [124] "ATL(R)"            "b1(R)"             "GF(R)"            
#> [127] "BU(L)"             "PB(L8)"            "PB"               
#> [130] "LO(R)"             "EPA(L)"            "PB(L4)"           
#> [133] "FB"                "SAD(-AMMC)"        "MB(R)"            
#> [136] "a'1(R)"            "SPS(L)"            "PB(L7)"           
#> [139] "CA(R)"             "aL(L)"             "SLP(R)"           
#> [142] "a1(R)"             "lACA(R)"           "b'L(L)"           
#> [145] "AOT(R)"            "bL(L)"             "g5(R)"            
#> [148] "PB(R3)"            "AB(L)"             "NO2(L)"           
#> [151] "WED(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method   path                                           
#> [1,] "GET"    "/api/cached/roicompleteness"                  
#> [2,] "POST"   "/api/v:ver/npexplorer/celltype/:dataset/:type"
#> [3,] "POST"   "/api/raw/cypher/transaction/:id/cypher"       
#> [4,] "DELETE" "/api/*"                                       
#> [5,] "TRACE"  "/api/*"                                       
#> [6,] "POST"   "/api/npexplorer/completeness"                 
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROICompleteness-fm"    
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [3,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [4,] "github.com/labstack/echo.glob..func1"                                                      
#> [5,] "github.com/labstack/echo.glob..func1"                                                      
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCompleteness-fm"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }