Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "b'1(R)"            "FBl4"              "aL(L)"            
#>   [4] "NO(R)"             "SAD(-AMMC)"        "FBl9"             
#>   [7] "PLP(R)"            "SIP(R)"            "EBr3am"           
#>  [10] "mALT(L)"           "SPS(R)"            "LX(R)"            
#>  [13] "FBl7"              "b'L(L)"            "GOR(L)"           
#>  [16] "PB(R2)"            "PB(R4)"            "PB"               
#>  [19] "SNP(R)"            "SNP(L)"            "FBl1"             
#>  [22] "NO(L)"             "g3(R)"             "LAL(-GA)(R)"      
#>  [25] "NO1(L)"            "PVLP(R)"           "LX(L)"            
#>  [28] "NO2(L)"            "b'L(R)"            "EBr5"             
#>  [31] "gL(R)"             "PB(R9)"            "ROB(R)"           
#>  [34] "EBr3pw"            "PB(R5)"            "PB(R7)"           
#>  [37] "MB(L)"             "PB(L1)"            "a'2(R)"           
#>  [40] "AL(L)"             "PB(L2)"            "GF(R)"            
#>  [43] "LAL(R)"            "SMP(R)"            "PENP"             
#>  [46] "NO3(L)"            "VES(R)"            "FBl8"             
#>  [49] "AMMC"              "a2(R)"             "hemibrain"        
#>  [52] "VMNP"              "MB(R)"             "EBr6"             
#>  [55] "EB"                "PB(R3)"            "dACA(R)"          
#>  [58] "EPA(L)"            "CRE(-ROB,-RUB)(R)" "PB(L9)"           
#>  [61] "g5(R)"             "a'L(L)"            "VLNP(R)"          
#>  [64] "AOT(R)"            "AOTU(R)"           "EBr1"             
#>  [67] "PB(L6)"            "PB(L5)"            "PED(R)"           
#>  [70] "LH(R)"             "OL(R)"             "AL(R)"            
#>  [73] "CA(R)"             "IPS(R)"            "GNG"              
#>  [76] "CAN(R)"            "SCL(L)"            "NO1(R)"           
#>  [79] "MB(+ACA)(R)"       "FBl2"              "SMP(L)"           
#>  [82] "g1(R)"             "EPA(R)"            "GA(R)"            
#>  [85] "ME(R)"             "IB"                "EBr3d"            
#>  [88] "PB(R1)"            "a'1(R)"            "GOR(R)"           
#>  [91] "PRW"               "SIP(L)"            "NO3(R)"           
#>  [94] "FBl6"              "ATL(R)"            "INP"              
#>  [97] "CA(L)"             "WED(R)"            "gL(L)"            
#> [100] "CRE(L)"            "ICL(R)"            "AB(L)"            
#> [103] "BU(R)"             "ICL(L)"            "PB(L4)"           
#> [106] "SLP(R)"            "CX"                "aL(R)"            
#> [109] "g2(R)"             "BU(L)"             "AME(R)"           
#> [112] "RUB(L)"            "SCL(R)"            "SPS(L)"           
#> [115] "AB(R)"             "b1(R)"             "a'3(R)"           
#> [118] "SAD"               "PB(L8)"            "a'L(R)"           
#> [121] "NO2(R)"            "POC"               "RUB(R)"           
#> [124] "PB(R6)"            "EBr2r4"            "bL(L)"            
#> [127] "g4(R)"             "bL(R)"             "a3(R)"            
#> [130] "CRE(R)"            "FB"                "FLA(R)"           
#> [133] "AVLP(R)"           "VES(L)"            "vACA(R)"          
#> [136] "a1(R)"             "lACA(R)"           "mALT(R)"          
#> [139] "FBl5"              "GC"                "b2(R)"            
#> [142] "PB(R8)"            "NO"                "PB(L7)"           
#> [145] "b'2(R)"            "LAL(L)"            "PB(L3)"           
#> [148] "ATL(L)"            "LO(R)"             "FBl3"             
#> [151] "LOP(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                      
#> [1,] "HEAD" "/api"                                    
#> [2,] "POST" "/api/npexplorer/simpleconnections"       
#> [3,] "GET"  "/api/v:ver/npexplorer/commonconnectivity"
#> [4,] "GET"  "/api/*"                                  
#> [5,] "POST" "/api/*"                                  
#> [6,] "GET"  "/api/v:ver/dbmeta/instances"             
#>      name                                                                                             
#> [1,] "github.com/labstack/echo.glob..func1"                                                           
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [4,] "github.com/labstack/echo.glob..func1"                                                           
#> [5,] "github.com/labstack/echo.glob..func1"                                                           
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }