Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "SAD"               "AL-DP1m(L)"        "AL-VA4(R)"        
#>   [4] "AL-VC1(R)"         "a'2(R)"            "AL-DL4(L)"        
#>   [7] "b1(R)"             "AL-DP1m(R)"        "IB"               
#>  [10] "CRE(-ROB,-RUB)(R)" "FBl1"              "FBl9"             
#>  [13] "FLA(R)"            "GOR(R)"            "LAL(L)"           
#>  [16] "PB(R5)"            "SMP(R)"            "OL(R)"            
#>  [19] "SPS(R)"            "SAD(-AMMC)"        "PB(L3)"           
#>  [22] "PLP(R)"            "AL-DA3(L)"         "AL-DM6(L)"        
#>  [25] "PB(L6)"            "SNP(L)"            "bL(R)"            
#>  [28] "AL-DM2(L)"         "AL-DC2(R)"         "EBr5"             
#>  [31] "EB"                "gL(L)"             "CAN(R)"           
#>  [34] "LOP(R)"            "aL(R)"             "AL-DC4(L)"        
#>  [37] "FBl8"              "PB(L2)"            "AL-DM1(L)"        
#>  [40] "FBl7"              "AL-VP5(R)"         "EBr1"             
#>  [43] "NO(R)"             "EBr3pw"            "AL-DA2(R)"        
#>  [46] "AL-DM2(R)"         "NO1(R)"            "AL-DC1(R)"        
#>  [49] "GF(R)"             "EBr6"              "AL-DM4(L)"        
#>  [52] "RUB(L)"            "AL-DC4(R)"         "AL-DM1(R)"        
#>  [55] "AL-VP1m(R)"        "AB(L)"             "hemibrain"        
#>  [58] "PB(R7)"            "FBl2"              "aL(L)"            
#>  [61] "GNG"               "AL-VL2a(R)"        "NO3(R)"           
#>  [64] "GOR(L)"            "bL(L)"             "AL-DL3(R)"        
#>  [67] "AL-VM3(R)"         "POC"               "ICL(L)"           
#>  [70] "PB(L7)"            "CA(R)"             "AL-DM3(R)"        
#>  [73] "AL-DP1l(R)"        "AL-VC4(R)"         "PB(L1)"           
#>  [76] "AL-VP1l(R)"        "VLNP(R)"           "WED(R)"           
#>  [79] "PB(R4)"            "AL-VA1d(R)"        "AL-VM7d(R)"       
#>  [82] "a'3(R)"            "SIP(L)"            "FB"               
#>  [85] "AL-DA2(L)"         "FB-column3"        "AL-DA4m(R)"       
#>  [88] "AL(L)"             "AL-VC5(R)"         "CX"               
#>  [91] "NO1(L)"            "BU(R)"             "NO"               
#>  [94] "AL-VA6(R)"         "AL-DM5(L)"         "PB(L9)"           
#>  [97] "LAL(R)"            "AL-DL1(R)"         "AL-VC3l(R)"       
#> [100] "BU(L)"             "PB(R8)"            "AL-DL2v(R)"       
#> [103] "AL-VC2(R)"         "SPS(L)"            "EBr2r4"           
#> [106] "ATL(R)"            "AL-DA1(R)"         "AL-VM5v(R)"       
#> [109] "SCL(L)"            "a'1(R)"            "mALT(L)"          
#> [112] "CRE(R)"            "SNP(R)"            "AL-VM7d(L)"       
#> [115] "NO3(L)"            "a'L(L)"            "AL-DA4l(R)"       
#> [118] "AL-DM6(R)"         "AL-D(R)"           "VMNP"             
#> [121] "g4(R)"             "VES(R)"            "AMMC"             
#> [124] "GA(R)"             "AL-DC1(L)"         "SIP(R)"           
#> [127] "PB(L4)"            "AL-DM4(R)"         "NO(L)"            
#> [130] "AL-DA3(R)"         "b'2(R)"            "SLP(R)"           
#> [133] "AL-VA3(R)"         "AL-VM5d(R)"        "AL-VP1d(R)"       
#> [136] "g5(R)"             "a1(R)"             "AL-VM7v(L)"       
#> [139] "AL-DL5(R)"         "AL-VP2(R)"         "PENP"             
#> [142] "a2(R)"             "MB(L)"             "AVLP(R)"          
#> [145] "MB(R)"             "AL-VA5(R)"         "IB112"            
#> [148] "PB(R2)"            "LX(L)"             "AL-DL2d(R)"       
#> [151] "dACA(R)"           "AL-VM2(R)"         "PRW"              
#> [154] "g2(R)"             "PVLP(R)"           "CA(L)"            
#> [157] "ICL(R)"            "AL-DM5(R)"         "PB(R6)"           
#> [160] "AL-VM1(R)"         "b'L(R)"            "AL-DL4(R)"        
#> [163] "AOT(R)"            "PED(R)"            "PB(L8)"           
#> [166] "ME(R)"             "AB(R)"             "a'L(R)"           
#> [169] "ROB(R)"            "NO2(R)"            "GC"               
#> [172] "LX(R)"             "INP"               "AL-VP4(R)"        
#> [175] "RUB(R)"            "PB(L5)"            "lACA(R)"          
#> [178] "AL-DC2(L)"         "AL-D(L)"           "FBl4"             
#> [181] "FBl6"              "PB(R3)"            "LO(R)"            
#> [184] "b'L(L)"            "SMP(L)"            "AL-VP3(R)"        
#> [187] "ATL(L)"            "g3(R)"             "AL-DA4m(L)"       
#> [190] "AL-VA6(L)"         "CRE(-RUB)(L)"      "AL-DC3(R)"        
#> [193] "AL-VA2(R)"         "AL-VM7v(R)"        "PB"               
#> [196] "AL-VA1v(R)"        "AL-VA7l(R)"        "AL-VA7m(R)"       
#> [199] "EBr3am"            "AOTU(R)"           "b2(R)"            
#> [202] "CRE(L)"            "mALT(R)"           "AL-DM3(L)"        
#> [205] "LH(R)"             "AME(R)"            "PB(R1)"           
#> [208] "FBl3"              "AL-VM4(R)"         "PB(R9)"           
#> [211] "EPA(L)"            "VES(L)"            "AL-VC3m(R)"       
#> [214] "AL-VL1(R)"         "AL-V(R)"           "FBl5"             
#> [217] "b'1(R)"            "IPS(R)"            "vACA(R)"          
#> [220] "MB(+ACA)(R)"       "a3(R)"             "NO2(L)"           
#> [223] "LAL(-GA)(R)"       "EBr3d"             "gL(R)"            
#> [226] "AL(R)"             "EPA(R)"            "AL-DL5(L)"        
#> [229] "g1(R)"             "SCL(R)"            "AL-VL2p(R)"       
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "EBr5"              "SIP(R)"            "ICL(R)"           
#>   [4] "PB(R6)"            "a2(R)"             "EPA(R)"           
#>   [7] "bL(L)"             "CA(R)"             "PENP"             
#>  [10] "GOR(L)"            "BU(L)"             "PLP(R)"           
#>  [13] "EBr3pw"            "MB(R)"             "WED(R)"           
#>  [16] "FB"                "dACA(R)"           "ME(R)"            
#>  [19] "bL(R)"             "AMMC"              "NO"               
#>  [22] "CRE(L)"            "IPS(R)"            "GOR(R)"           
#>  [25] "LOP(R)"            "aL(R)"             "LX(R)"            
#>  [28] "FBl4"              "RUB(L)"            "RUB(R)"           
#>  [31] "PB(L3)"            "FLA(R)"            "SPS(R)"           
#>  [34] "aL(L)"             "gL(R)"             "NO3(L)"           
#>  [37] "PB(L2)"            "SCL(R)"            "PB(R9)"           
#>  [40] "FBl8"              "EPA(L)"            "GNG"              
#>  [43] "lACA(R)"           "PB(R7)"            "mALT(L)"          
#>  [46] "b'2(R)"            "ICL(L)"            "FBl3"             
#>  [49] "PB(R3)"            "SMP(R)"            "SLP(R)"           
#>  [52] "POC"               "a'3(R)"            "GF(R)"            
#>  [55] "SIP(L)"            "b2(R)"             "PB(L6)"           
#>  [58] "SAD(-AMMC)"        "NO(R)"             "VES(R)"           
#>  [61] "SCL(L)"            "PRW"               "LAL(-GA)(R)"      
#>  [64] "PB(R8)"            "VMNP"              "BU(R)"            
#>  [67] "a'L(L)"            "g3(R)"             "g1(R)"            
#>  [70] "EBr6"              "MB(+ACA)(R)"       "FBl1"             
#>  [73] "AOTU(R)"           "NO1(R)"            "SNP(L)"           
#>  [76] "hemibrain"         "SAD"               "VES(L)"           
#>  [79] "PB(R5)"            "AB(R)"             "g2(R)"            
#>  [82] "EBr1"              "a'2(R)"            "g5(R)"            
#>  [85] "LO(R)"             "gL(L)"             "PB(L9)"           
#>  [88] "FBl6"              "NO2(R)"            "FBl9"             
#>  [91] "AME(R)"            "vACA(R)"           "NO3(R)"           
#>  [94] "PB(L8)"            "ROB(R)"            "PVLP(R)"          
#>  [97] "FBl2"              "PED(R)"            "FBl7"             
#> [100] "OL(R)"             "mALT(R)"           "PB(R4)"           
#> [103] "VLNP(R)"           "AOT(R)"            "LX(L)"            
#> [106] "SMP(L)"            "a'L(R)"            "a'1(R)"           
#> [109] "CRE(-ROB,-RUB)(R)" "a1(R)"             "b1(R)"            
#> [112] "NO(L)"             "AB(L)"             "MB(L)"            
#> [115] "PB(L1)"            "g4(R)"             "PB(L7)"           
#> [118] "PB(R2)"            "CRE(R)"            "b'1(R)"           
#> [121] "CA(L)"             "PB(L4)"            "GA(R)"            
#> [124] "SPS(L)"            "EBr2r4"            "AVLP(R)"          
#> [127] "b'L(L)"            "a3(R)"             "IB"               
#> [130] "SNP(R)"            "AL(L)"             "PB(R1)"           
#> [133] "LH(R)"             "CAN(R)"            "NO1(L)"           
#> [136] "b'L(R)"            "INP"               "FBl5"             
#> [139] "LAL(R)"            "CX"                "NO2(L)"           
#> [142] "LAL(L)"            "EBr3d"             "PB(L5)"           
#> [145] "ATL(L)"            "EBr3am"            "GC"               
#> [148] "PB"                "EB"                "ATL(R)"           
#> [151] "AL(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "GF(R)"             "AL-DM4(L)"         "AL-DL5(R)"        
#>   [4] "AL-D(R)"           "SPS(R)"            "dACA(R)"          
#>   [7] "PB(R3)"            "FBl9"              "AL-DA4m(R)"       
#>  [10] "vACA(R)"           "ME(R)"             "gL(L)"            
#>  [13] "PB(L9)"            "EBr3pw"            "PB(R4)"           
#>  [16] "AL-DL1(R)"         "AL-DM2(R)"         "AL-VC5(R)"        
#>  [19] "b'L(R)"            "EBr3d"             "AL(L)"            
#>  [22] "FBl8"              "b'1(R)"            "AL-DA3(R)"        
#>  [25] "AL-V(R)"           "AL-VP5(R)"         "SPS(L)"           
#>  [28] "a'L(R)"            "NO3(L)"            "EPA(L)"           
#>  [31] "bL(L)"             "g5(R)"             "AL-DM5(R)"        
#>  [34] "AL-DP1l(R)"        "g4(R)"             "SLP(R)"           
#>  [37] "AL-VP1m(R)"        "AL-VP1d(R)"        "CRE(-ROB,-RUB)(R)"
#>  [40] "AMMC"              "b'2(R)"            "GOR(L)"           
#>  [43] "FBl5"              "b'L(L)"            "CRE(-RUB)(L)"     
#>  [46] "AL-DM4(R)"         "CA(L)"             "b2(R)"            
#>  [49] "PB(L6)"            "bL(R)"             "PLP(R)"           
#>  [52] "AL-DA1(R)"         "a1(R)"             "a'L(L)"           
#>  [55] "AL-DL3(R)"         "PB(R5)"            "ATL(L)"           
#>  [58] "AL-DL4(L)"         "AL-DL2v(R)"        "AL-DL4(R)"        
#>  [61] "AL-VP3(R)"         "AL-DM2(L)"         "FB-column3"       
#>  [64] "AL-DA2(R)"         "AL-VM7d(R)"        "AL-VA5(R)"        
#>  [67] "PENP"              "SCL(R)"            "AB(R)"            
#>  [70] "LX(L)"             "AL-VM7v(L)"        "NO2(R)"           
#>  [73] "AME(R)"            "INP"               "GOR(R)"           
#>  [76] "SIP(L)"            "g1(R)"             "EBr5"             
#>  [79] "LH(R)"             "AL-DM3(R)"         "AL-DM6(R)"        
#>  [82] "EBr2r4"            "AL-DL5(L)"         "ROB(R)"           
#>  [85] "CRE(L)"            "AL-VA1v(R)"        "AL-VL2p(R)"       
#>  [88] "GA(R)"             "g2(R)"             "PB(R9)"           
#>  [91] "SAD(-AMMC)"        "PB"                "AL-DM5(L)"        
#>  [94] "ATL(R)"            "AL-VA7l(R)"        "AL-DC1(R)"        
#>  [97] "AL-VA7m(R)"        "AL-VM7v(R)"        "SAD"              
#> [100] "EBr6"              "PB(L8)"            "a'2(R)"           
#> [103] "LAL(L)"            "AL-DC2(R)"         "AL-VM5v(R)"       
#> [106] "g3(R)"             "PVLP(R)"           "CX"               
#> [109] "PB(L7)"            "RUB(L)"            "AL-VA2(R)"        
#> [112] "AB(L)"             "PB(L3)"            "FBl2"             
#> [115] "AL-VM3(R)"         "SMP(R)"            "hemibrain"        
#> [118] "LAL(-GA)(R)"       "EPA(R)"            "NO"               
#> [121] "ICL(L)"            "AL-DC4(R)"         "PED(R)"           
#> [124] "AL-DM3(L)"         "AL-VA6(L)"         "NO(R)"            
#> [127] "CAN(R)"            "AL-VC3l(R)"        "NO3(R)"           
#> [130] "VMNP"              "AL-VM7d(L)"        "MB(+ACA)(R)"      
#> [133] "GC"                "AL-DL2d(R)"        "IB"               
#> [136] "a'3(R)"            "FBl1"              "POC"              
#> [139] "FBl3"              "GNG"               "AL-VL1(R)"        
#> [142] "AL-VM1(R)"         "AL-VA3(R)"         "RUB(R)"           
#> [145] "SCL(L)"            "WED(R)"            "PB(R8)"           
#> [148] "EB"                "AL-DC4(L)"         "AOTU(R)"          
#> [151] "mALT(R)"           "OL(R)"             "AL-VP2(R)"        
#> [154] "AL-DA3(L)"         "a'1(R)"            "BU(L)"            
#> [157] "EBr3am"            "PB(R6)"            "NO(L)"            
#> [160] "aL(R)"             "EBr1"              "FLA(R)"           
#> [163] "AL-VM4(R)"         "PB(L4)"            "VES(L)"           
#> [166] "VLNP(R)"           "AL-D(L)"           "PB(R1)"           
#> [169] "AOT(R)"            "SNP(R)"            "CRE(R)"           
#> [172] "IPS(R)"            "LAL(R)"            "AL-VP4(R)"        
#> [175] "MB(L)"             "PB(L1)"            "aL(L)"            
#> [178] "AL-VL2a(R)"        "PRW"               "FBl6"             
#> [181] "AL-DP1m(R)"        "AL-DP1m(L)"        "SNP(L)"           
#> [184] "FBl4"              "a3(R)"             "lACA(R)"          
#> [187] "NO1(L)"            "BU(R)"             "AL-DM1(L)"        
#> [190] "AL-DM1(R)"         "LOP(R)"            "gL(R)"            
#> [193] "b1(R)"             "SMP(L)"            "LX(R)"            
#> [196] "AVLP(R)"           "AL-DC1(L)"         "AL-VA1d(R)"       
#> [199] "AL-DA4l(R)"        "AL-VC2(R)"         "PB(L2)"           
#> [202] "NO1(R)"            "AL-DM6(L)"         "AL-VC3m(R)"       
#> [205] "LO(R)"             "PB(R2)"            "MB(R)"            
#> [208] "CA(R)"             "AL-VP1l(R)"        "AL-DA4m(L)"       
#> [211] "PB(L5)"            "AL-VA4(R)"         "SIP(R)"           
#> [214] "FBl7"              "AL-VM2(R)"         "AL-DA2(L)"        
#> [217] "ICL(R)"            "AL-VA6(R)"         "mALT(L)"          
#> [220] "AL-DC2(L)"         "VES(R)"            "AL-VC4(R)"        
#> [223] "AL-VM5d(R)"        "AL(R)"             "AL-VC1(R)"        
#> [226] "NO2(L)"            "a2(R)"             "PB(R7)"           
#> [229] "FB"                "AL-DC3(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "PLP(R)"            "AL-VC4(R)"         "AL-VL1(R)"        
#>   [4] "AL-VL2a(R)"        "PB(L2)"            "AMMC"             
#>   [7] "AL-DL3(R)"         "AL-VC2(R)"         "lACA(R)"          
#>  [10] "AL-DL5(L)"         "AL-DA3(R)"         "GNG"              
#>  [13] "AL-DM3(R)"         "AL-VA1d(R)"        "b2(R)"            
#>  [16] "VMNP"              "CRE(-RUB)(L)"      "FBl4"             
#>  [19] "AB(R)"             "AL-VM7v(R)"        "SMP(L)"           
#>  [22] "a'L(L)"            "AL-VM1(R)"         "PB(L7)"           
#>  [25] "PVLP(R)"           "NO3(R)"            "AL-DM5(L)"        
#>  [28] "AL(R)"             "AL-D(R)"           "NO2(L)"           
#>  [31] "a2(R)"             "LAL(-GA)(R)"       "ICL(L)"           
#>  [34] "NO1(R)"            "OL(R)"             "AL-VP1d(R)"       
#>  [37] "AL-VP1l(R)"        "PB(R5)"            "PED(R)"           
#>  [40] "GOR(L)"            "AL-DP1m(L)"        "aL(L)"            
#>  [43] "AL-DM6(R)"         "mALT(R)"           "ME(R)"            
#>  [46] "AVLP(R)"           "AL-VC3m(R)"        "IPS(R)"           
#>  [49] "AME(R)"            "RUB(L)"            "AL-DL1(R)"        
#>  [52] "AL-VP2(R)"         "g2(R)"             "g4(R)"            
#>  [55] "b'1(R)"            "PB"                "AL-VA6(L)"        
#>  [58] "NO"                "VES(L)"            "MB(R)"            
#>  [61] "AL-DA1(R)"         "vACA(R)"           "g5(R)"            
#>  [64] "FBl9"              "CA(R)"             "LX(R)"            
#>  [67] "AL-VM7d(L)"        "PB(R9)"            "PB(L6)"           
#>  [70] "ROB(R)"            "VES(R)"            "PB(L4)"           
#>  [73] "AL-VC5(R)"         "PB(L8)"            "PB(R1)"           
#>  [76] "FLA(R)"            "PB(L1)"            "a'1(R)"           
#>  [79] "MB(L)"             "EB"                "PB(L9)"           
#>  [82] "AL-DL4(R)"         "AL-DM4(L)"         "EBr3pw"           
#>  [85] "SPS(R)"            "AL-DL2v(R)"        "b'L(R)"           
#>  [88] "AL-DL4(L)"         "PB(R4)"            "AL-VM7d(R)"       
#>  [91] "AL-VA7m(R)"        "EBr3d"             "AL-DC1(L)"        
#>  [94] "FBl1"              "RUB(R)"            "SNP(R)"           
#>  [97] "FB"                "AL-DC2(R)"         "PB(R8)"           
#> [100] "MB(+ACA)(R)"       "AL-DC4(R)"         "AL-VM4(R)"        
#> [103] "FBl2"              "FBl5"              "FBl6"             
#> [106] "NO2(R)"            "AL-DC1(R)"         "AL-VA4(R)"        
#> [109] "AL-DM2(L)"         "AL-VA3(R)"         "AL-VA5(R)"        
#> [112] "WED(R)"            "dACA(R)"           "AL-DM3(L)"        
#> [115] "FB-column3"        "AL-DA2(R)"         "AL-DP1l(R)"       
#> [118] "AOTU(R)"           "AL-DC2(L)"         "b1(R)"            
#> [121] "AL-VA6(R)"         "AL-VC1(R)"         "SPS(L)"           
#> [124] "GOR(R)"            "CX"                "GA(R)"            
#> [127] "a'2(R)"            "EPA(R)"            "AL-D(L)"          
#> [130] "AL-VA7l(R)"        "AOT(R)"            "g3(R)"            
#> [133] "EBr1"              "FBl8"              "AL-DM4(R)"        
#> [136] "ATL(L)"            "ICL(R)"            "AL-DC3(R)"        
#> [139] "SCL(L)"            "PB(L3)"            "SIP(L)"           
#> [142] "NO1(L)"            "AL-DC4(L)"         "gL(L)"            
#> [145] "b'L(L)"            "SLP(R)"            "AL-V(R)"          
#> [148] "GC"                "gL(R)"             "SAD(-AMMC)"       
#> [151] "INP"               "AL-VP4(R)"         "AB(L)"            
#> [154] "a'3(R)"            "LX(L)"             "LOP(R)"           
#> [157] "mALT(L)"           "GF(R)"             "aL(R)"            
#> [160] "BU(L)"             "BU(R)"             "VLNP(R)"          
#> [163] "POC"               "a'L(R)"            "LO(R)"            
#> [166] "NO(L)"             "AL-DM1(R)"         "SAD"              
#> [169] "PB(R3)"            "AL-VP5(R)"         "PB(L5)"           
#> [172] "CRE(R)"            "AL-DM6(L)"         "AL-DA4l(R)"       
#> [175] "IB"                "a1(R)"             "AL-VP3(R)"        
#> [178] "AL-VP1m(R)"        "PB(R6)"            "EBr5"             
#> [181] "CRE(L)"            "AL-DA4m(R)"        "AL-VM2(R)"        
#> [184] "bL(L)"             "FBl3"              "AL-VM5v(R)"       
#> [187] "CA(L)"             "SIP(R)"            "AL-DL2d(R)"       
#> [190] "AL-DP1m(R)"        "AL-VA1v(R)"        "AL-VM5d(R)"       
#> [193] "PB(R2)"            "EBr3am"            "PRW"              
#> [196] "AL-DM5(R)"         "AL-VL2p(R)"        "LAL(L)"           
#> [199] "SCL(R)"            "SNP(L)"            "a3(R)"            
#> [202] "AL-DA2(L)"         "LAL(R)"            "AL-VA2(R)"        
#> [205] "g1(R)"             "AL-VM7v(L)"        "LH(R)"            
#> [208] "CAN(R)"            "AL-DA3(L)"         "AL-DA4m(L)"       
#> [211] "AL-DM1(L)"         "NO3(L)"            "EPA(L)"           
#> [214] "AL-DL5(R)"         "AL-DM2(R)"         "AL-VM3(R)"        
#> [217] "FBl7"              "bL(R)"             "PB(R7)"           
#> [220] "NO(R)"             "ATL(R)"            "AL-VC3l(R)"       
#> [223] "PENP"              "b'2(R)"            "hemibrain"        
#> [226] "EBr6"              "AL(L)"             "EBr2r4"           
#> [229] "SMP(R)"            "CRE(-ROB,-RUB)(R)"
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                     
#> [1,] "GET"  "/api/v:ver/npexplorer/autapses"         
#> [2,] "GET"  "/api/v:ver/roimeshes/mesh/:dataset/:roi"
#> [3,] "POST" "/api/v:ver/roimeshes/mesh/:dataset/:roi"
#> [4,] "POST" "/api/v:ver/raw/cypher/cypher"           
#> [5,] "GET"  "/api/npexplorer/commonconnectivity"     
#> [6,] "GET"  "/api/roimeshes/mesh/:dataset/:roi"      
#>      name                                                                                              
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCommonConnectivity-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/roimeshes.masterAPI.getMesh-fm"                  

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }