Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "PB(R7)"            "AL-VM7v(L)"        "AL-VM7v(R)"       
#>   [4] "LAL(-GA)(R)"       "NO1(L)"            "AL-DL2v(R)"       
#>   [7] "AL-VA3(R)"         "AL-VM4(R)"         "AL-DL5(L)"        
#>  [10] "AL-DM5(L)"         "NO(L)"             "AL-VP4(R)"        
#>  [13] "hemibrain"         "NO1(R)"            "AL-D(L)"          
#>  [16] "NO3(R)"            "VES(L)"            "FBl4"             
#>  [19] "PB(R3)"            "aL(L)"             "AL-VP1m(R)"       
#>  [22] "PB(R6)"            "ICL(L)"            "VES(R)"           
#>  [25] "LH(R)"             "CX"                "INP"              
#>  [28] "AL-DC3(R)"         "AL-VC3l(R)"        "WED(R)"           
#>  [31] "AL-VP1d(R)"        "CA(R)"             "AL-VC1(R)"        
#>  [34] "AOTU(R)"           "GA(R)"             "EBr3d"            
#>  [37] "NO2(R)"            "a3(R)"             "AL-VA1d(R)"       
#>  [40] "AL-VM1(R)"         "EB"                "FBl7"             
#>  [43] "AB(R)"             "AL-VM7d(L)"        "AL-VC3m(R)"       
#>  [46] "IB"                "FB"                "EBr6"             
#>  [49] "g4(R)"             "SPS(R)"            "AL-DM6(R)"        
#>  [52] "AL-VP2(R)"         "MB(+ACA)(R)"       "NO2(L)"           
#>  [55] "SCL(R)"            "gL(L)"             "LAL(R)"           
#>  [58] "EPA(L)"            "AL-DA4l(R)"        "SCL(L)"           
#>  [61] "PLP(R)"            "PVLP(R)"           "LX(L)"            
#>  [64] "AL-DP1l(R)"        "AL-VA2(R)"         "AOT(R)"           
#>  [67] "b'1(R)"            "LOP(R)"            "AL-DM6(L)"        
#>  [70] "FBl1"              "AL-DM2(R)"         "AL-DL2d(R)"       
#>  [73] "AL(R)"             "AL-DL3(R)"         "POC"              
#>  [76] "aL(R)"             "dACA(R)"           "AL-DM2(L)"        
#>  [79] "CA(L)"             "b'2(R)"            "CRE(L)"           
#>  [82] "AL-VP5(R)"         "AL-DM4(R)"         "a2(R)"            
#>  [85] "EBr5"              "AL-DL1(R)"         "CRE(-ROB,-RUB)(R)"
#>  [88] "SNP(R)"            "AVLP(R)"           "AL-VL1(R)"        
#>  [91] "AL-DC4(L)"         "bL(R)"             "IB112"            
#>  [94] "MB(R)"             "AL-VC5(R)"         "GOR(L)"           
#>  [97] "AL(L)"             "AL-DA2(L)"         "AB(L)"            
#> [100] "CRE(R)"            "EBr1"              "BU(R)"            
#> [103] "MB(L)"             "OL(R)"             "gL(R)"            
#> [106] "CRE(-RUB)(L)"      "AL-V(R)"           "PENP"             
#> [109] "SAD"               "g2(R)"             "AL-DA4m(R)"       
#> [112] "AL-VA7m(R)"        "PB(L3)"            "AL-DC1(L)"        
#> [115] "PB(R9)"            "AL-VM3(R)"         "AL-VP3(R)"        
#> [118] "PB(L8)"            "PB(R8)"            "EBr3pw"           
#> [121] "AL-VC2(R)"         "FLA(R)"            "AL-DM1(L)"        
#> [124] "SAD(-AMMC)"        "AL-DP1m(R)"        "mALT(L)"          
#> [127] "PED(R)"            "a'1(R)"            "FBl5"             
#> [130] "EBr3am"            "RUB(L)"            "AL-VL2p(R)"       
#> [133] "AL-VM7d(R)"        "BU(L)"             "AL-VM2(R)"        
#> [136] "LAL(L)"            "SIP(R)"            "AL-VA4(R)"        
#> [139] "g3(R)"             "EBr2r4"            "AL-DA2(R)"        
#> [142] "AL-DA4m(L)"        "AL-VM5v(R)"        "AL-DA1(R)"        
#> [145] "AL-DA3(R)"         "b'L(R)"            "PB(R1)"           
#> [148] "ATL(R)"            "GNG"               "AL-VA6(R)"        
#> [151] "g1(R)"             "AME(R)"            "PB(R4)"           
#> [154] "FBl9"              "LX(R)"             "SNP(L)"           
#> [157] "FBl8"              "g5(R)"             "AL-DL5(R)"        
#> [160] "AL-VA7l(R)"        "a'3(R)"            "AL-DA3(L)"        
#> [163] "NO"                "AL-DL4(R)"         "AL-VA5(R)"        
#> [166] "PB(L5)"            "AL-VL2a(R)"        "VLNP(R)"          
#> [169] "GF(R)"             "AL-VA6(L)"         "IPS(R)"           
#> [172] "mALT(R)"           "SMP(R)"            "AL-DC2(L)"        
#> [175] "NO(R)"             "PB(R5)"            "AL-VM5d(R)"       
#> [178] "SIP(L)"            "AL-DC2(R)"         "AL-DC4(R)"        
#> [181] "AL-VC4(R)"         "a1(R)"             "AL-DM3(R)"        
#> [184] "PRW"               "FBl2"              "FBl3"             
#> [187] "LO(R)"             "PB(L1)"            "PB"               
#> [190] "ME(R)"             "AL-DM3(L)"         "PB(L9)"           
#> [193] "PB(R2)"            "PB(L4)"            "GC"               
#> [196] "a'L(R)"            "ROB(R)"            "EPA(R)"           
#> [199] "AL-DP1m(L)"        "GOR(R)"            "AL-VA1v(R)"       
#> [202] "RUB(R)"            "SMP(L)"            "VMNP"             
#> [205] "a'2(R)"            "PB(L2)"            "ICL(R)"           
#> [208] "AL-D(R)"           "b2(R)"             "AL-DM4(L)"        
#> [211] "NO3(L)"            "FB-column3"        "b'L(L)"           
#> [214] "vACA(R)"           "SPS(L)"            "AL-DL4(L)"        
#> [217] "bL(L)"             "PB(L6)"            "AL-DM5(R)"        
#> [220] "PB(L7)"            "AL-VP1l(R)"        "b1(R)"            
#> [223] "AL-DC1(R)"         "ATL(L)"            "FBl6"             
#> [226] "CAN(R)"            "AL-DM1(R)"         "SLP(R)"           
#> [229] "lACA(R)"           "a'L(L)"            "AMMC"             
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "EBr3pw"            "LO(R)"             "PB(R5)"           
#>   [4] "a'3(R)"            "PB(R6)"            "WED(R)"           
#>   [7] "g2(R)"             "EBr1"              "ME(R)"            
#>  [10] "b1(R)"             "IB"                "FBl1"             
#>  [13] "PB(R3)"            "EBr5"              "SNP(R)"           
#>  [16] "a'1(R)"            "FBl9"              "a1(R)"            
#>  [19] "SCL(R)"            "EB"                "g5(R)"            
#>  [22] "a3(R)"             "AB(L)"             "PB(L5)"           
#>  [25] "CX"                "LAL(L)"            "PRW"              
#>  [28] "EPA(R)"            "bL(L)"             "CA(R)"            
#>  [31] "CA(L)"             "NO2(L)"            "MB(L)"            
#>  [34] "g4(R)"             "LX(L)"             "PB"               
#>  [37] "AL(L)"             "gL(R)"             "SNP(L)"           
#>  [40] "VES(L)"            "AL(R)"             "SPS(R)"           
#>  [43] "EBr6"              "ICL(L)"            "PB(L7)"           
#>  [46] "FBl7"              "LAL(-GA)(R)"       "b2(R)"            
#>  [49] "BU(R)"             "MB(+ACA)(R)"       "FBl3"             
#>  [52] "ICL(R)"            "PB(L8)"            "PLP(R)"           
#>  [55] "a'L(L)"            "SMP(R)"            "AMMC"             
#>  [58] "hemibrain"         "SAD"               "NO1(R)"           
#>  [61] "VMNP"              "NO"                "CRE(L)"           
#>  [64] "POC"               "g3(R)"             "LH(R)"            
#>  [67] "b'1(R)"            "FLA(R)"            "a'2(R)"           
#>  [70] "SIP(R)"            "NO2(R)"            "EBr3am"           
#>  [73] "GF(R)"             "PB(R8)"            "AOT(R)"           
#>  [76] "PB(L2)"            "ATL(L)"            "BU(L)"            
#>  [79] "NO(R)"             "SAD(-AMMC)"        "EPA(L)"           
#>  [82] "SCL(L)"            "aL(R)"             "CAN(R)"           
#>  [85] "GOR(R)"            "OL(R)"             "ATL(R)"           
#>  [88] "AB(R)"             "SLP(R)"            "mALT(L)"          
#>  [91] "SMP(L)"            "PB(L1)"            "LX(R)"            
#>  [94] "FBl8"              "PB(L4)"            "NO3(R)"           
#>  [97] "ROB(R)"            "CRE(-ROB,-RUB)(R)" "GOR(L)"           
#> [100] "GC"                "SPS(L)"            "PB(R9)"           
#> [103] "VLNP(R)"           "LOP(R)"            "EBr2r4"           
#> [106] "a'L(R)"            "FBl5"              "INP"              
#> [109] "PB(R4)"            "NO(L)"             "NO1(L)"           
#> [112] "b'L(R)"            "AME(R)"            "vACA(R)"          
#> [115] "b'2(R)"            "aL(L)"             "PENP"             
#> [118] "AOTU(R)"           "GA(R)"             "PED(R)"           
#> [121] "mALT(R)"           "NO3(L)"            "VES(R)"           
#> [124] "GNG"               "LAL(R)"            "lACA(R)"          
#> [127] "CRE(R)"            "FB"                "FBl2"             
#> [130] "AVLP(R)"           "PB(L3)"            "a2(R)"            
#> [133] "PB(R7)"            "gL(L)"             "PB(R1)"           
#> [136] "PB(L9)"            "bL(R)"             "b'L(L)"           
#> [139] "MB(R)"             "RUB(R)"            "EBr3d"            
#> [142] "PVLP(R)"           "g1(R)"             "dACA(R)"          
#> [145] "PB(L6)"            "FBl4"              "SIP(L)"           
#> [148] "FBl6"              "PB(R2)"            "IPS(R)"           
#> [151] "RUB(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DL5(R)"         "AL-VP1m(R)"        "CX"               
#>   [4] "NO1(R)"            "PB(R8)"            "AL-DM3(L)"        
#>   [7] "RUB(L)"            "LOP(R)"            "GC"               
#>  [10] "dACA(R)"           "AL-DM2(L)"         "VES(L)"           
#>  [13] "AL-DP1l(R)"        "AL-VA7m(R)"        "NO1(L)"           
#>  [16] "AL-VM4(R)"         "AL-VP1d(R)"        "PLP(R)"           
#>  [19] "AL-VA1d(R)"        "a1(R)"             "AL-D(L)"          
#>  [22] "AL-DP1m(L)"        "PB(R2)"            "AL-DA4m(R)"       
#>  [25] "PB(R6)"            "AVLP(R)"           "AL-DA2(R)"        
#>  [28] "SCL(L)"            "PB(L3)"            "a'2(R)"           
#>  [31] "INP"               "AL(R)"             "AL-VM5v(R)"       
#>  [34] "mALT(L)"           "PB"                "FBl2"             
#>  [37] "OL(R)"             "PB(L6)"            "PB(L9)"           
#>  [40] "bL(L)"             "AL-VM5d(R)"        "b'2(R)"           
#>  [43] "a'1(R)"            "LH(R)"             "FBl1"             
#>  [46] "lACA(R)"           "ME(R)"             "LX(L)"            
#>  [49] "AL-VL2p(R)"        "b1(R)"             "AL-VA4(R)"        
#>  [52] "SIP(L)"            "AL-DC1(L)"         "ICL(L)"           
#>  [55] "LO(R)"             "AL-VL1(R)"         "SPS(R)"           
#>  [58] "AL-DA3(L)"         "AL-DA4m(L)"        "SNP(R)"           
#>  [61] "MB(+ACA)(R)"       "AL-VA5(R)"         "FBl9"             
#>  [64] "ICL(R)"            "SMP(R)"            "NO2(L)"           
#>  [67] "PB(L1)"            "FLA(R)"            "CRE(-RUB)(L)"     
#>  [70] "EBr3pw"            "NO2(R)"            "AL-DM2(R)"        
#>  [73] "b'L(R)"            "AL-DM5(L)"         "AL-VA6(L)"        
#>  [76] "bL(R)"             "EBr3d"             "ROB(R)"           
#>  [79] "AL-DL4(R)"         "PB(L2)"            "GOR(L)"           
#>  [82] "CRE(-ROB,-RUB)(R)" "CA(L)"             "a'L(L)"           
#>  [85] "AL-VM7d(R)"        "PENP"              "a'L(R)"           
#>  [88] "AL-DC4(L)"         "AL-DM1(L)"         "PB(L7)"           
#>  [91] "FBl7"              "AL-V(R)"           "b'1(R)"           
#>  [94] "PRW"               "AL-DM4(L)"         "SNP(L)"           
#>  [97] "AL-VA1v(R)"        "AL-DL4(L)"         "AL-DM4(R)"        
#> [100] "SIP(R)"            "FBl5"              "POC"              
#> [103] "LX(R)"             "AL-VM7v(L)"        "aL(L)"            
#> [106] "PB(L5)"            "AL-DC4(R)"         "EBr5"             
#> [109] "vACA(R)"           "FB"                "g4(R)"            
#> [112] "PB(R1)"            "FBl6"              "AL-VC2(R)"        
#> [115] "AL-VC1(R)"         "AL-VL2a(R)"        "AL-VM2(R)"        
#> [118] "AL-VM7v(R)"        "VLNP(R)"           "AL-VP3(R)"        
#> [121] "LAL(L)"            "aL(R)"             "BU(L)"            
#> [124] "VES(R)"            "b'L(L)"            "SMP(L)"           
#> [127] "NO3(R)"            "AL-DC2(L)"         "MB(R)"            
#> [130] "LAL(-GA)(R)"       "NO(L)"             "AL-DA1(R)"        
#> [133] "EBr3am"            "a2(R)"             "PB(L8)"           
#> [136] "AL-DM3(R)"         "AL-DP1m(R)"        "AL-VC4(R)"        
#> [139] "CRE(R)"            "g2(R)"             "CRE(L)"           
#> [142] "FBl8"              "AL-VC3l(R)"        "NO(R)"            
#> [145] "AL-VP4(R)"         "AB(L)"             "AMMC"             
#> [148] "CAN(R)"            "SAD"               "WED(R)"           
#> [151] "SCL(R)"            "SAD(-AMMC)"        "FBl3"             
#> [154] "IPS(R)"            "gL(R)"             "CA(R)"            
#> [157] "AL-DL1(R)"         "AL-VA6(R)"         "AL-VC5(R)"        
#> [160] "NO3(L)"            "PB(R3)"            "EPA(L)"           
#> [163] "AL-DC1(R)"         "AOTU(R)"           "PB(R9)"           
#> [166] "AL-VA2(R)"         "EBr1"              "AL-VM1(R)"        
#> [169] "VMNP"              "PB(L4)"            "AL-DL2v(R)"       
#> [172] "AL-VC3m(R)"        "ATL(L)"            "GF(R)"            
#> [175] "PED(R)"            "AB(R)"             "GA(R)"            
#> [178] "PVLP(R)"           "AL-VM3(R)"         "AL-DM5(R)"        
#> [181] "EBr6"              "SPS(L)"            "AL-DA4l(R)"       
#> [184] "g5(R)"             "IB"                "AL-DA2(L)"        
#> [187] "AL-DL2d(R)"        "AL-DL3(R)"         "gL(L)"            
#> [190] "AL-DM6(L)"         "FBl4"              "AL-D(R)"          
#> [193] "FB-column3"        "AME(R)"            "AL-DC3(R)"        
#> [196] "hemibrain"         "mALT(R)"           "EB"               
#> [199] "AL-VA3(R)"         "MB(L)"             "NO"               
#> [202] "SLP(R)"            "EPA(R)"            "GOR(R)"           
#> [205] "AL-DC2(R)"         "AOT(R)"            "AL-DL5(L)"        
#> [208] "AL-VM7d(L)"        "AL(L)"             "AL-DA3(R)"        
#> [211] "AL-DM1(R)"         "AL-DM6(R)"         "AL-VP1l(R)"       
#> [214] "a3(R)"             "a'3(R)"            "AL-VP5(R)"        
#> [217] "g1(R)"             "EBr2r4"            "BU(R)"            
#> [220] "GNG"               "LAL(R)"            "AL-VA7l(R)"       
#> [223] "b2(R)"             "PB(R4)"            "PB(R5)"           
#> [226] "AL-VP2(R)"         "RUB(R)"            "g3(R)"            
#> [229] "PB(R7)"            "ATL(R)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "a'2(R)"            "EBr3pw"            "SNP(L)"           
#>   [4] "AL-DL2d(R)"        "AL-VC3l(R)"        "AL-VP4(R)"        
#>   [7] "a'L(R)"            "EPA(R)"            "INP"              
#>  [10] "AL-DP1m(R)"        "PB(R7)"            "NO"               
#>  [13] "AL-VP1m(R)"        "PB(L5)"            "AL-DM3(L)"        
#>  [16] "SNP(R)"            "AME(R)"            "AL-VA1v(R)"       
#>  [19] "NO3(R)"            "FLA(R)"            "PB(R1)"           
#>  [22] "FBl8"              "AL-DA1(R)"         "AL-DC4(R)"        
#>  [25] "AL-DM3(R)"         "PB(L1)"            "ICL(L)"           
#>  [28] "AL-VM5v(R)"        "AL-VM7v(R)"        "GA(R)"            
#>  [31] "MB(L)"             "SCL(R)"            "VES(L)"           
#>  [34] "PB(L2)"            "gL(L)"             "AL-DC1(R)"        
#>  [37] "AL-VM1(R)"         "AL-VP3(R)"         "AB(L)"            
#>  [40] "WED(R)"            "PB(L7)"            "PB(L4)"           
#>  [43] "AL-DL5(R)"         "PB(L8)"            "ATL(R)"           
#>  [46] "AL-DA4l(R)"        "PB(R9)"            "FBl1"             
#>  [49] "AL-DL3(R)"         "b2(R)"             "VLNP(R)"          
#>  [52] "b'1(R)"            "ATL(L)"            "a2(R)"            
#>  [55] "AL-VA3(R)"         "BU(R)"             "LOP(R)"           
#>  [58] "b'2(R)"            "AL-VC4(R)"         "PRW"              
#>  [61] "PVLP(R)"           "NO2(R)"            "AL-VA2(R)"        
#>  [64] "AL-VA7m(R)"        "lACA(R)"           "EBr3d"            
#>  [67] "b1(R)"             "a'L(L)"            "AL-VA6(R)"        
#>  [70] "AL-VM2(R)"         "bL(R)"             "AL-VM5d(R)"       
#>  [73] "NO2(L)"            "FBl7"              "aL(L)"            
#>  [76] "AL-VA4(R)"         "AL-VM3(R)"         "ME(R)"            
#>  [79] "PLP(R)"            "EBr3am"            "EBr6"             
#>  [82] "GOR(R)"            "AL-VP1d(R)"        "SCL(L)"           
#>  [85] "AL-DM1(L)"         "PB(L6)"            "AMMC"             
#>  [88] "AL-VM7d(L)"        "AL-VC2(R)"         "AVLP(R)"          
#>  [91] "AL-DL1(R)"         "RUB(R)"            "AL-D(R)"          
#>  [94] "AL-VA5(R)"         "PB(R5)"            "vACA(R)"          
#>  [97] "BU(L)"             "g4(R)"             "AL-DM5(L)"        
#> [100] "AL-DC2(R)"         "AL(R)"             "AL-DM1(R)"        
#> [103] "NO1(R)"            "PB"                "ROB(R)"           
#> [106] "CAN(R)"            "AL-DC4(L)"         "AL-VL1(R)"        
#> [109] "AL-DM2(R)"         "AL-DP1l(R)"        "SMP(L)"           
#> [112] "AL-VM7v(L)"        "FB-column3"        "LH(R)"            
#> [115] "a3(R)"             "LAL(R)"            "AL-DA3(R)"        
#> [118] "AL-VP2(R)"         "SAD"               "GC"               
#> [121] "g1(R)"             "CRE(L)"            "AL-VP1l(R)"       
#> [124] "a1(R)"             "PB(R8)"            "EBr2r4"           
#> [127] "AL-DA3(L)"         "AL-DC2(L)"         "RUB(L)"           
#> [130] "AL-DM6(R)"         "NO(L)"             "GF(R)"            
#> [133] "NO(R)"             "FBl3"              "GNG"              
#> [136] "AL-DC3(R)"         "AOT(R)"            "a'1(R)"           
#> [139] "IB"                "AL-DM6(L)"         "SAD(-AMMC)"       
#> [142] "FBl5"              "AL-VM7d(R)"        "AL-VP5(R)"        
#> [145] "PB(R6)"            "aL(R)"             "CRE(R)"           
#> [148] "FBl2"              "SPS(L)"            "NO3(L)"           
#> [151] "AL-VC5(R)"         "mALT(R)"           "AL-V(R)"          
#> [154] "CA(R)"             "VMNP"              "bL(L)"            
#> [157] "PB(L9)"            "IPS(R)"            "SMP(R)"           
#> [160] "AL-DM2(L)"         "MB(R)"             "AL-VC1(R)"        
#> [163] "PB(R3)"            "AL-DA2(R)"         "SPS(R)"           
#> [166] "GOR(L)"            "EB"                "AL-DA4m(L)"       
#> [169] "LAL(L)"            "b'L(L)"            "AL-DM5(R)"        
#> [172] "CX"                "AL-VA6(L)"         "SLP(R)"           
#> [175] "PB(L3)"            "AL(L)"             "AL-D(L)"          
#> [178] "AL-DM4(L)"         "AL-VA7l(R)"        "AL-VL2p(R)"       
#> [181] "dACA(R)"           "PB(R2)"            "LX(L)"            
#> [184] "AL-VA1d(R)"        "LO(R)"             "EBr1"             
#> [187] "ICL(R)"            "AL-DA2(L)"         "AB(R)"            
#> [190] "AL-DL2v(R)"        "POC"               "hemibrain"        
#> [193] "VES(R)"            "FBl4"              "FBl6"             
#> [196] "g3(R)"             "NO1(L)"            "FBl9"             
#> [199] "a'3(R)"            "PENP"              "g2(R)"            
#> [202] "EPA(L)"            "CRE(-ROB,-RUB)(R)" "mALT(L)"          
#> [205] "CA(L)"             "PED(R)"            "AL-DL5(L)"        
#> [208] "SIP(R)"            "AL-VM4(R)"         "EBr5"             
#> [211] "AL-DP1m(L)"        "g5(R)"             "PB(R4)"           
#> [214] "AL-DM4(R)"         "b'L(R)"            "CRE(-RUB)(L)"     
#> [217] "MB(+ACA)(R)"       "AL-DL4(R)"         "AOTU(R)"          
#> [220] "SIP(L)"            "AL-DC1(L)"         "gL(R)"            
#> [223] "AL-VC3m(R)"        "AL-VL2a(R)"        "FB"               
#> [226] "LX(R)"             "OL(R)"             "AL-DA4m(R)"       
#> [229] "LAL(-GA)(R)"       "AL-DL4(L)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                        
#> [1,] "PUT"  "/api"                                      
#> [2,] "POST" "/api/v:ver/npexplorer/rankedtable"         
#> [3,] "POST" "/api/v:ver/npexplorer/roisinneuron"        
#> [4,] "GET"  "/api/npexplorer/completeness"              
#> [5,] "GET"  "/api/v:ver/cached/roicompleteness"         
#> [6,] "POST" "/api/v:ver/raw/keyvalue/key/:instance/:key"
#>      name                                                                                        
#> [1,] "github.com/labstack/echo.glob..func1"                                                      
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCompleteness-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }