Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "b2(R)"             "AL-DC2(R)"         "hemibrain"        
#>   [4] "AL-DM5(L)"         "NO3(L)"            "FB-column3"       
#>   [7] "AL-VM3(R)"         "b'L(R)"            "AL-DA2(L)"        
#>  [10] "SNP(L)"            "LH(R)"             "ICL(R)"           
#>  [13] "AL-VA7m(R)"        "VLNP(R)"           "ATL(L)"           
#>  [16] "NO1(L)"            "AL-DM6(L)"         "AL-VM7d(L)"       
#>  [19] "PB(L9)"            "PB(L7)"            "bL(R)"            
#>  [22] "AL-VM5v(R)"        "SMP(L)"            "EBr5"             
#>  [25] "PB(L2)"            "SAD"               "g1(R)"            
#>  [28] "BU(L)"             "LX(R)"             "AL-DA4m(L)"       
#>  [31] "AL-DL4(L)"         "AL-DM3(L)"         "CA(R)"            
#>  [34] "a'2(R)"            "EBr3d"             "AB(R)"            
#>  [37] "AMMC"              "PB(R9)"            "PB(L4)"           
#>  [40] "AL-VL2a(R)"        "EBr3pw"            "AL-VA2(R)"        
#>  [43] "AL-VC1(R)"         "IB"                "PB(R6)"           
#>  [46] "g2(R)"             "PB"                "AL-DC2(L)"        
#>  [49] "gL(R)"             "MB(L)"             "IB112"            
#>  [52] "FBl5"              "PB(R3)"            "AL-DA4l(R)"       
#>  [55] "AL-VM2(R)"         "AL-DP1m(L)"        "AL-VP5(R)"        
#>  [58] "AL-DC4(L)"         "AL-DL4(R)"         "AL-VC2(R)"        
#>  [61] "AL-VM4(R)"         "FBl6"              "CX"               
#>  [64] "PB(L6)"            "AL-VP1d(R)"        "GNG"              
#>  [67] "b'L(L)"            "OL(R)"             "AL-DL2d(R)"       
#>  [70] "AL-VA1v(R)"        "AL-V(R)"           "AL-VP1m(R)"       
#>  [73] "SCL(L)"            "SIP(L)"            "PB(L1)"           
#>  [76] "BU(R)"             "AL-VM7v(L)"        "AL-DL3(R)"        
#>  [79] "SIP(R)"            "AL-DA1(R)"         "GOR(L)"           
#>  [82] "SNP(R)"            "a'L(R)"            "a3(R)"            
#>  [85] "AL-VA3(R)"         "PRW"               "AL-DP1l(R)"       
#>  [88] "PB(R7)"            "PB(L8)"            "a2(R)"            
#>  [91] "SPS(R)"            "AL-DM2(R)"         "NO3(R)"           
#>  [94] "EBr2r4"            "AL-DL5(L)"         "MB(+ACA)(R)"      
#>  [97] "AB(L)"             "PED(R)"            "AL(L)"            
#> [100] "ROB(R)"            "AL-VA7l(R)"        "AOT(R)"           
#> [103] "CRE(-ROB,-RUB)(R)" "g4(R)"             "AL-VC3l(R)"       
#> [106] "LAL(-GA)(R)"       "PLP(R)"            "AL-VA4(R)"        
#> [109] "AL-VA5(R)"         "b'2(R)"            "NO1(R)"           
#> [112] "AL-DA3(L)"         "AL-DA4m(R)"        "AL-VP3(R)"        
#> [115] "CA(L)"             "AL-DC1(L)"         "GOR(R)"           
#> [118] "AL-VM1(R)"         "GF(R)"             "NO(R)"            
#> [121] "RUB(L)"            "a'3(R)"            "NO2(L)"           
#> [124] "LAL(L)"            "GC"                "LX(L)"            
#> [127] "AL-DL2v(R)"        "CRE(R)"            "EBr1"             
#> [130] "AL-VM5d(R)"        "PENP"              "AL-VA6(L)"        
#> [133] "FBl9"              "AME(R)"            "aL(L)"            
#> [136] "a1(R)"             "SLP(R)"            "AL-DC4(R)"        
#> [139] "AL-DM1(R)"         "LOP(R)"            "PB(R1)"           
#> [142] "AL-DL1(R)"         "AL-DM3(R)"         "AL-DM4(R)"        
#> [145] "GA(R)"             "aL(R)"             "NO2(R)"           
#> [148] "PB(R2)"            "PB(L5)"            "mALT(L)"          
#> [151] "FLA(R)"            "FBl8"              "FBl3"             
#> [154] "AVLP(R)"           "mALT(R)"           "FB"               
#> [157] "PB(R4)"            "PB(R5)"            "PB(L3)"           
#> [160] "g5(R)"             "AL-DL5(R)"         "AL-VP1l(R)"       
#> [163] "b'1(R)"            "lACA(R)"           "EBr6"             
#> [166] "AL-D(L)"           "FBl4"              "AL-DP1m(R)"       
#> [169] "SCL(R)"            "PVLP(R)"           "AL-VM7d(R)"       
#> [172] "bL(L)"             "FBl7"              "AL-DM5(R)"        
#> [175] "AL-DM4(L)"         "LAL(R)"            "AL-VP2(R)"        
#> [178] "vACA(R)"           "g3(R)"             "EPA(L)"           
#> [181] "MB(R)"             "AL-DA2(R)"         "AL-DM6(R)"        
#> [184] "AL-VC3m(R)"        "SMP(R)"            "LO(R)"            
#> [187] "AL(R)"             "CRE(L)"            "EPA(R)"           
#> [190] "AL-DC3(R)"         "AL-D(R)"           "AL-VC4(R)"        
#> [193] "AL-VP4(R)"         "FBl2"              "dACA(R)"          
#> [196] "AL-DM1(L)"         "NO"                "AL-VA6(R)"        
#> [199] "VMNP"              "gL(L)"             "FBl1"             
#> [202] "b1(R)"             "VES(R)"            "ATL(R)"           
#> [205] "AL-VL2p(R)"        "POC"               "WED(R)"           
#> [208] "EB"                "IPS(R)"            "NO(L)"            
#> [211] "AL-DA3(R)"         "AL-VC5(R)"         "RUB(R)"           
#> [214] "AL-DM2(L)"         "AL-VM7v(R)"        "SAD(-AMMC)"       
#> [217] "VES(L)"            "AL-VL1(R)"         "AOTU(R)"          
#> [220] "SPS(L)"            "PB(R8)"            "ME(R)"            
#> [223] "CRE(-RUB)(L)"      "a'1(R)"            "CAN(R)"           
#> [226] "INP"               "AL-DC1(R)"         "AL-VA1d(R)"       
#> [229] "ICL(L)"            "a'L(L)"            "EBr3am"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "VES(L)"            "AB(L)"             "vACA(R)"          
#>   [4] "PB(R9)"            "NO(L)"             "a'3(R)"           
#>   [7] "WED(R)"            "MB(L)"             "LAL(R)"           
#>  [10] "AOTU(R)"           "b'L(L)"            "LO(R)"            
#>  [13] "FBl4"              "OL(R)"             "a2(R)"            
#>  [16] "LX(R)"             "FBl3"              "VLNP(R)"          
#>  [19] "AOT(R)"            "SCL(L)"            "b'L(R)"           
#>  [22] "FLA(R)"            "PB(R7)"            "PB(L8)"           
#>  [25] "LX(L)"             "g3(R)"             "SIP(L)"           
#>  [28] "a'2(R)"            "MB(+ACA)(R)"       "FBl9"             
#>  [31] "IPS(R)"            "GF(R)"             "mALT(R)"          
#>  [34] "NO1(R)"            "SIP(R)"            "SAD(-AMMC)"       
#>  [37] "INP"               "AL(R)"             "LOP(R)"           
#>  [40] "SMP(L)"            "NO3(R)"            "PB(L5)"           
#>  [43] "EBr3d"             "FBl5"              "a3(R)"            
#>  [46] "IB"                "GOR(L)"            "PB"               
#>  [49] "FBl2"              "LH(R)"             "PB(L4)"           
#>  [52] "CRE(R)"            "PB(R1)"            "CRE(-ROB,-RUB)(R)"
#>  [55] "ICL(L)"            "aL(L)"             "CRE(L)"           
#>  [58] "RUB(L)"            "MB(R)"             "b'2(R)"           
#>  [61] "FB"                "PB(R8)"            "gL(L)"            
#>  [64] "g5(R)"             "SAD"               "PED(R)"           
#>  [67] "bL(L)"             "EPA(L)"            "aL(R)"            
#>  [70] "EBr5"              "FBl8"              "AVLP(R)"          
#>  [73] "lACA(R)"           "GC"                "g2(R)"            
#>  [76] "BU(L)"             "EBr3pw"            "VES(R)"           
#>  [79] "b1(R)"             "GOR(R)"            "POC"              
#>  [82] "SNP(R)"            "NO(R)"             "EPA(R)"           
#>  [85] "NO"                "PB(R2)"            "AL(L)"            
#>  [88] "PB(L9)"            "EBr3am"            "mALT(L)"          
#>  [91] "ATL(L)"            "GA(R)"             "PB(L7)"           
#>  [94] "EBr1"              "SLP(R)"            "RUB(R)"           
#>  [97] "PB(L2)"            "PENP"              "SPS(R)"           
#> [100] "a'L(R)"            "FBl6"              "ATL(R)"           
#> [103] "PB(R3)"            "a1(R)"             "PB(L3)"           
#> [106] "hemibrain"         "NO2(L)"            "SNP(L)"           
#> [109] "bL(R)"             "GNG"               "SMP(R)"           
#> [112] "g1(R)"             "NO3(L)"            "ICL(R)"           
#> [115] "PB(R6)"            "SPS(L)"            "VMNP"             
#> [118] "AMMC"              "NO1(L)"            "PB(L6)"           
#> [121] "BU(R)"             "a'1(R)"            "CA(R)"            
#> [124] "AB(R)"             "a'L(L)"            "CX"               
#> [127] "EBr6"              "ME(R)"             "CA(L)"            
#> [130] "AME(R)"            "LAL(-GA)(R)"       "SCL(R)"           
#> [133] "gL(R)"             "FBl1"              "FBl7"             
#> [136] "NO2(R)"            "PB(L1)"            "g4(R)"            
#> [139] "PLP(R)"            "EBr2r4"            "EB"               
#> [142] "PVLP(R)"           "PRW"               "dACA(R)"          
#> [145] "ROB(R)"            "CAN(R)"            "PB(R4)"           
#> [148] "PB(R5)"            "b'1(R)"            "LAL(L)"           
#> [151] "b2(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "MB(L)"             "EB"                "AL-DL2d(R)"       
#>   [4] "AL-DC4(L)"         "AL-DL1(R)"         "SPS(L)"           
#>   [7] "AL-DM4(R)"         "AL-DM5(R)"         "NO3(R)"           
#>  [10] "LX(R)"             "AL(L)"             "AB(R)"            
#>  [13] "AL-VC3l(R)"        "SMP(R)"            "AL-VP1l(R)"       
#>  [16] "ME(R)"             "AL-D(L)"           "PB(L9)"           
#>  [19] "g5(R)"             "AL-VM3(R)"         "AL-VP4(R)"        
#>  [22] "g1(R)"             "RUB(L)"            "b'L(L)"           
#>  [25] "AL(R)"             "RUB(R)"            "AL-DC2(L)"        
#>  [28] "AL-DP1m(L)"        "AL-VP5(R)"         "AL-DM3(L)"        
#>  [31] "PB(L3)"            "AL-DA3(L)"         "AL-DL4(L)"        
#>  [34] "AL-VA6(L)"         "AL-VA1v(R)"        "VLNP(R)"          
#>  [37] "SAD"               "NO1(R)"            "PB(R1)"           
#>  [40] "AL-VP2(R)"         "POC"               "IB"               
#>  [43] "AL-VA5(R)"         "LAL(L)"            "CRE(-RUB)(L)"     
#>  [46] "VES(R)"            "vACA(R)"           "b'2(R)"           
#>  [49] "b2(R)"             "AL-VM4(R)"         "AL-V(R)"          
#>  [52] "PENP"              "PB(R8)"            "AL-DC1(L)"        
#>  [55] "FBl5"              "PB(L2)"            "b'L(R)"           
#>  [58] "gL(L)"             "AL-DA3(R)"         "AL-VC4(R)"        
#>  [61] "PB(R6)"            "PRW"               "g2(R)"            
#>  [64] "FBl4"              "NO2(L)"            "PB(R7)"           
#>  [67] "PB(L8)"            "MB(+ACA)(R)"       "LAL(R)"           
#>  [70] "AL-VA7l(R)"        "LOP(R)"            "aL(R)"            
#>  [73] "CA(R)"             "a'3(R)"            "AL-DA1(R)"        
#>  [76] "AL-DC3(R)"         "AL-DC4(R)"         "SNP(R)"           
#>  [79] "AL-DM2(L)"         "SIP(R)"            "AL-VA3(R)"        
#>  [82] "GF(R)"             "g4(R)"             "AOT(R)"           
#>  [85] "EPA(R)"            "AL-DA2(L)"         "ROB(R)"           
#>  [88] "AL-VA4(R)"         "AL-VM1(R)"         "IPS(R)"           
#>  [91] "AL-VM5d(R)"        "AL-VM7v(R)"        "AL-VP3(R)"        
#>  [94] "FLA(R)"            "PB(R9)"            "FBl1"             
#>  [97] "EPA(L)"            "AL-DL3(R)"         "AL-VM7d(R)"       
#> [100] "g3(R)"             "NO1(L)"            "BU(R)"            
#> [103] "LO(R)"             "PVLP(R)"           "LH(R)"            
#> [106] "PB(L5)"            "CA(L)"             "FB"               
#> [109] "NO2(R)"            "lACA(R)"           "CRE(R)"           
#> [112] "EBr5"              "aL(L)"             "AL-VA1d(R)"       
#> [115] "AL-VC2(R)"         "AL-DM6(R)"         "dACA(R)"          
#> [118] "PB(L1)"            "EBr3d"             "bL(L)"            
#> [121] "FBl9"              "ATL(R)"            "GOR(L)"           
#> [124] "SAD(-AMMC)"        "FBl2"              "PB(R4)"           
#> [127] "NO3(L)"            "AL-DA2(R)"         "AL-DM2(R)"        
#> [130] "AL-VM5v(R)"        "SCL(L)"            "AOTU(R)"          
#> [133] "EBr3pw"            "GOR(R)"            "CRE(-ROB,-RUB)(R)"
#> [136] "PB(L6)"            "AL-VA2(R)"         "GA(R)"            
#> [139] "FB-column3"        "b1(R)"             "AL-DA4l(R)"       
#> [142] "AL-VP1m(R)"        "b'1(R)"            "SMP(L)"           
#> [145] "MB(R)"             "AL-D(R)"           "a1(R)"            
#> [148] "OL(R)"             "AL-DL4(R)"         "AL-VA7m(R)"       
#> [151] "AL-VL2p(R)"        "AL-VP1d(R)"        "AL-DM1(L)"        
#> [154] "AL-VC1(R)"         "AL-VC5(R)"         "GC"               
#> [157] "AL-DM1(R)"         "SCL(R)"            "PB(L4)"           
#> [160] "LX(L)"             "AL-VC3m(R)"        "CRE(L)"           
#> [163] "AL-DC2(R)"         "AL-VM2(R)"         "GNG"              
#> [166] "a'L(L)"            "AL-DP1m(R)"        "a'2(R)"           
#> [169] "AL-DA4m(L)"        "gL(R)"             "PB(L7)"           
#> [172] "bL(R)"             "AL-DA4m(R)"        "ATL(L)"           
#> [175] "CAN(R)"            "VES(L)"            "AL-DM3(R)"        
#> [178] "SPS(R)"            "EBr2r4"            "SNP(L)"           
#> [181] "FBl3"              "FBl6"              "AME(R)"           
#> [184] "PED(R)"            "ICL(R)"            "a3(R)"            
#> [187] "AL-VA6(R)"         "EBr6"              "AL-DM4(L)"        
#> [190] "PB(R2)"            "SIP(L)"            "WED(R)"           
#> [193] "VMNP"              "NO(R)"             "INP"              
#> [196] "AL-VL1(R)"         "BU(L)"             "AL-DM6(L)"        
#> [199] "AL-VM7d(L)"        "PB(R3)"            "EBr3am"           
#> [202] "AL-DL5(L)"         "mALT(L)"           "AMMC"             
#> [205] "a2(R)"             "hemibrain"         "PB"               
#> [208] "PLP(R)"            "AL-DM5(L)"         "AL-DP1l(R)"       
#> [211] "AL-VM7v(L)"        "a'1(R)"            "SLP(R)"           
#> [214] "CX"                "a'L(R)"            "AB(L)"            
#> [217] "AVLP(R)"           "AL-VL2a(R)"        "LAL(-GA)(R)"      
#> [220] "FBl8"              "AL-DC1(R)"         "AL-DL2v(R)"       
#> [223] "mALT(R)"           "NO"                "ICL(L)"           
#> [226] "FBl7"              "NO(L)"             "EBr1"             
#> [229] "AL-DL5(R)"         "PB(R5)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "g1(R)"             "FB-column3"        "EBr3pw"           
#>   [4] "AL-VC2(R)"         "EBr6"              "VES(R)"           
#>   [7] "FBl9"              "AL-VA6(R)"         "EBr1"             
#>  [10] "EBr2r4"            "EB"                "NO2(L)"           
#>  [13] "GC"                "b2(R)"             "PB(L1)"           
#>  [16] "PB(L5)"            "AMMC"              "FLA(R)"           
#>  [19] "AL-DA3(L)"         "PB(R1)"            "NO2(R)"           
#>  [22] "aL(L)"             "INP"               "EBr3am"           
#>  [25] "AL-DC4(L)"         "PB(L6)"            "a'1(R)"           
#>  [28] "GNG"               "AL-VC4(R)"         "b'L(L)"           
#>  [31] "AL-DA4l(R)"        "AL-DM3(R)"         "AL-V(R)"          
#>  [34] "mALT(L)"           "hemibrain"         "SPS(L)"           
#>  [37] "AL-DP1m(L)"        "EPA(L)"            "SCL(L)"           
#>  [40] "PB(L2)"            "PB(R7)"            "PLP(R)"           
#>  [43] "ME(R)"             "EBr3d"             "AL-VM7v(L)"       
#>  [46] "NO(R)"             "LH(R)"             "AL-VM1(R)"        
#>  [49] "AL-DA4m(L)"        "AL-DC2(L)"         "AL-DL4(L)"        
#>  [52] "NO"                "AL-DC1(R)"         "AL-DL5(R)"        
#>  [55] "AL-VP3(R)"         "b'1(R)"            "PRW"              
#>  [58] "gL(R)"             "PB(R2)"            "PB(L8)"           
#>  [61] "AL-VA7l(R)"        "AL-VM7d(R)"        "GF(R)"            
#>  [64] "BU(L)"             "ATL(L)"            "AL-DL5(L)"        
#>  [67] "AL-VM7d(L)"        "CA(R)"             "ICL(L)"           
#>  [70] "AL-VA3(R)"         "VMNP"              "SNP(R)"           
#>  [73] "g4(R)"             "AL-DM4(R)"         "AL-VM7v(R)"       
#>  [76] "SMP(R)"            "FBl2"              "AL-DP1m(R)"       
#>  [79] "AL-VC5(R)"         "LX(R)"             "AL-DM6(L)"        
#>  [82] "AL-DM6(R)"         "AL-VA1v(R)"        "AL-VC3l(R)"       
#>  [85] "AL-VP1l(R)"        "CRE(-ROB,-RUB)(R)" "a2(R)"            
#>  [88] "AL-DL2v(R)"        "AL-VM4(R)"         "AOTU(R)"          
#>  [91] "FB"                "OL(R)"             "PENP"             
#>  [94] "WED(R)"            "AL-VM2(R)"         "AOT(R)"           
#>  [97] "LO(R)"             "AL-VP5(R)"         "AL-VP1m(R)"       
#> [100] "GOR(L)"            "b'L(R)"            "NO1(R)"           
#> [103] "gL(L)"             "IPS(R)"            "AL-VP4(R)"        
#> [106] "g2(R)"             "AL-D(L)"           "AL-D(R)"          
#> [109] "AL-VA1d(R)"        "AL-VC3m(R)"        "AL-VA5(R)"        
#> [112] "VLNP(R)"           "SIP(R)"            "ROB(R)"           
#> [115] "AL-DM2(R)"         "AB(L)"             "AL-DC3(R)"        
#> [118] "AL-VP2(R)"         "AL-DM5(L)"         "FBl1"             
#> [121] "FBl4"              "AL-VM5v(R)"        "POC"              
#> [124] "PB(R5)"            "a'2(R)"            "AL(L)"            
#> [127] "AL-DC4(R)"         "AL-VC1(R)"         "a'3(R)"           
#> [130] "aL(R)"             "EPA(R)"            "AL-DA2(L)"        
#> [133] "PB(R4)"            "vACA(R)"           "a'L(R)"           
#> [136] "bL(L)"             "PVLP(R)"           "AVLP(R)"          
#> [139] "AL-DM5(R)"         "AL-DL3(R)"         "AL-VL2p(R)"       
#> [142] "PB(R6)"            "g3(R)"             "AL-DM1(L)"        
#> [145] "MB(+ACA)(R)"       "GOR(R)"            "MB(R)"            
#> [148] "b'2(R)"            "NO3(R)"            "FBl6"             
#> [151] "AL-DL2d(R)"        "LOP(R)"            "ATL(R)"           
#> [154] "AL-DC2(R)"         "CA(L)"             "SMP(L)"           
#> [157] "CRE(L)"            "AL-DA4m(R)"        "AL-VA2(R)"        
#> [160] "CX"                "AL-VP1d(R)"        "SIP(L)"           
#> [163] "PB(R3)"            "IB"                "lACA(R)"          
#> [166] "PB(R8)"            "AL(R)"             "SCL(R)"           
#> [169] "AL-VL2a(R)"        "AL-VM5d(R)"        "PB(L3)"           
#> [172] "GA(R)"             "SAD"               "AL-DC1(L)"        
#> [175] "NO3(L)"            "bL(R)"             "a'L(L)"           
#> [178] "AL-DL1(R)"         "AL-DL4(R)"         "SPS(R)"           
#> [181] "AL-DM3(L)"         "ICL(R)"            "SLP(R)"           
#> [184] "AL-DA1(R)"         "a1(R)"             "SNP(L)"           
#> [187] "FBl5"              "a3(R)"             "AL-DM1(R)"        
#> [190] "PB(L4)"            "AL-DA2(R)"         "AL-DP1l(R)"       
#> [193] "PED(R)"            "CRE(-RUB)(L)"      "g5(R)"            
#> [196] "FBl3"              "AB(R)"             "MB(L)"            
#> [199] "EBr5"              "LAL(L)"            "dACA(R)"          
#> [202] "b1(R)"             "LAL(R)"            "BU(R)"            
#> [205] "LX(L)"             "AL-VA6(L)"         "SAD(-AMMC)"       
#> [208] "AL-DM2(L)"         "PB(L7)"            "AME(R)"           
#> [211] "RUB(R)"            "mALT(R)"           "CAN(R)"           
#> [214] "RUB(L)"            "LAL(-GA)(R)"       "CRE(R)"           
#> [217] "FBl8"              "NO(L)"             "AL-VA4(R)"        
#> [220] "NO1(L)"            "PB(R9)"            "FBl7"             
#> [223] "AL-DA3(R)"         "VES(L)"            "AL-VA7m(R)"       
#> [226] "AL-VL1(R)"         "PB"                "AL-DM4(L)"        
#> [229] "PB(L9)"            "AL-VM3(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method   path                                
#> [1,] "HEAD"   "/api"                              
#> [2,] "GET"    "/api/v:ver/npexplorer/roisinneuron"
#> [3,] "GET"    "/api/npexplorer/autapses"          
#> [4,] "DELETE" "/api"                              
#> [5,] "GET"    "/api/custom/custom"                
#> [6,] "GET"    "/oauth2callback"                   
#>      name                                                                                    
#> [1,] "github.com/labstack/echo.glob..func1"                                                  
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"
#> [4,] "github.com/labstack/echo.glob..func1"                                                  
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"         
#> [6,] "github.com/janelia-flyem/echo-secure.oauthCallbackHandler"                             
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }