Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DC1(R)"         "AL-VM2(R)"         "b'L(L)"           
#>   [4] "AL-DA4l(R)"        "AL-DM6(R)"         "AL-DP1l(R)"       
#>   [7] "AL-DM2(L)"         "CAN(R)"            "ROB(R)"           
#>  [10] "LH(R)"             "CRE(-ROB,-RUB)(R)" "PED(R)"           
#>  [13] "AL-DL1(R)"         "AL-VP4(R)"         "g2(R)"            
#>  [16] "FB-column3"        "AL-VM1(R)"         "a'2(R)"           
#>  [19] "LX(L)"             "EBr2r4"            "OL(R)"            
#>  [22] "PB(R9)"            "SAD(-AMMC)"        "PENP"             
#>  [25] "dACA(R)"           "NO1(R)"            "PB(R8)"           
#>  [28] "PB(R1)"            "AL-VM5d(R)"        "NO1(L)"           
#>  [31] "BU(L)"             "AL-VA6(L)"         "AL-VP2(R)"        
#>  [34] "SPS(R)"            "PB(L4)"            "FBl8"             
#>  [37] "AMMC"              "AL-DM6(L)"         "AL-V(R)"          
#>  [40] "AL-DA2(L)"         "POC"               "CA(R)"            
#>  [43] "a'L(L)"            "AL-DA4m(R)"        "VLNP(R)"          
#>  [46] "PB(L8)"            "ME(R)"             "AL-VL1(R)"        
#>  [49] "NO3(R)"            "VES(R)"            "AL-VA4(R)"        
#>  [52] "vACA(R)"           "AL-DL3(R)"         "AL-VM3(R)"        
#>  [55] "AOTU(R)"           "FB"                "AL-VP1m(R)"       
#>  [58] "bL(R)"             "AL-VC4(R)"         "AL-DL4(R)"        
#>  [61] "lACA(R)"           "BU(R)"             "SPS(L)"           
#>  [64] "PB(R5)"            "SNP(R)"            "a'L(R)"           
#>  [67] "FBl4"              "LAL(L)"            "g3(R)"            
#>  [70] "PB(R2)"            "GNG"               "LAL(R)"           
#>  [73] "b'L(R)"            "EB"                "AL-DL4(L)"        
#>  [76] "AL-VA6(R)"         "CA(L)"             "GA(R)"            
#>  [79] "AL-DC1(L)"         "LX(R)"             "MB(R)"            
#>  [82] "AL-VC3l(R)"        "FBl2"              "AL-VA1d(R)"       
#>  [85] "SIP(R)"            "EPA(L)"            "EBr3d"            
#>  [88] "AL-VP5(R)"         "IPS(R)"            "AL-DM1(R)"        
#>  [91] "AL-VC3m(R)"        "SIP(L)"            "AL-DC4(L)"        
#>  [94] "NO2(R)"            "AL-VA2(R)"         "a1(R)"            
#>  [97] "FBl3"              "AL-VP1d(R)"        "RUB(R)"           
#> [100] "b'1(R)"            "FBl9"              "AL-VP1l(R)"       
#> [103] "PB(R3)"            "AL-VM7v(L)"        "MB(+ACA)(R)"      
#> [106] "LAL(-GA)(R)"       "AL-VC5(R)"         "ATL(L)"           
#> [109] "AL-DC4(R)"         "AL-VA7m(R)"        "ICL(R)"           
#> [112] "NO(L)"             "PB(R7)"            "EBr3pw"           
#> [115] "AL-VM5v(R)"        "EBr3am"            "AL-DM2(R)"        
#> [118] "AL-D(R)"           "AL-DL2v(R)"        "a2(R)"            
#> [121] "PB"                "FBl6"              "AL-VM7d(R)"       
#> [124] "LOP(R)"            "CX"                "WED(R)"           
#> [127] "CRE(-RUB)(L)"      "AL(L)"             "a'1(R)"           
#> [130] "IB112"             "AL-DA2(R)"         "AL-DP1m(R)"       
#> [133] "g1(R)"             "PB(L1)"            "AL-DA1(R)"        
#> [136] "AL-VP3(R)"         "AL-DP1m(L)"        "FBl1"             
#> [139] "NO2(L)"            "PB(L3)"            "AL-DM4(L)"        
#> [142] "g5(R)"             "RUB(L)"            "PB(R4)"           
#> [145] "MB(L)"             "NO"                "GOR(R)"           
#> [148] "gL(R)"             "PB(L7)"            "AL-VM7v(R)"       
#> [151] "SMP(R)"            "a'3(R)"            "VMNP"             
#> [154] "AL-VC1(R)"         "SCL(R)"            "EBr1"             
#> [157] "EBr5"              "gL(L)"             "AL-DA4m(L)"       
#> [160] "LO(R)"             "PRW"               "mALT(R)"          
#> [163] "AL-VM4(R)"         "FBl7"              "AME(R)"           
#> [166] "b1(R)"             "AL-VL2p(R)"        "EPA(R)"           
#> [169] "AL-DM1(L)"         "aL(R)"             "NO3(L)"           
#> [172] "AL-VA7l(R)"        "PB(L2)"            "b'2(R)"           
#> [175] "GC"                "AL-DL5(R)"         "PB(L9)"           
#> [178] "FLA(R)"            "AL-DL5(L)"         "AL(R)"            
#> [181] "AL-VL2a(R)"        "AL-DM5(L)"         "a3(R)"            
#> [184] "mALT(L)"           "NO(R)"             "PB(L5)"           
#> [187] "CRE(R)"            "AL-DM3(R)"         "AL-DM4(R)"        
#> [190] "AL-VC2(R)"         "PB(R6)"            "SMP(L)"           
#> [193] "AL-DC3(R)"         "AL-DL2d(R)"        "g4(R)"            
#> [196] "PVLP(R)"           "SLP(R)"            "AL-DA3(L)"        
#> [199] "VES(L)"            "AL-VM7d(L)"        "ICL(L)"           
#> [202] "FBl5"              "aL(L)"             "INP"              
#> [205] "AL-VA3(R)"         "GF(R)"             "SAD"              
#> [208] "AB(L)"             "CRE(L)"            "AL-DA3(R)"        
#> [211] "hemibrain"         "PB(L6)"            "SNP(L)"           
#> [214] "ATL(R)"            "AL-VA1v(R)"        "AL-VA5(R)"        
#> [217] "AOT(R)"            "b2(R)"             "AL-DM3(L)"        
#> [220] "AL-D(L)"           "SCL(L)"            "AVLP(R)"          
#> [223] "AB(R)"             "AL-DC2(R)"         "AL-DM5(R)"        
#> [226] "IB"                "PLP(R)"            "bL(L)"            
#> [229] "AL-DC2(L)"         "GOR(L)"            "EBr6"             
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "LX(R)"             "FBl2"              "PB(R3)"           
#>   [4] "PB(R4)"            "bL(R)"             "AB(L)"            
#>   [7] "a'3(R)"            "MB(L)"             "FBl5"             
#>  [10] "GC"                "a'L(R)"            "PB(L9)"           
#>  [13] "FBl4"              "ICL(L)"            "LH(R)"            
#>  [16] "b'L(L)"            "AL(R)"             "b'2(R)"           
#>  [19] "lACA(R)"           "EBr5"              "dACA(R)"          
#>  [22] "AME(R)"            "a'1(R)"            "NO3(L)"           
#>  [25] "CA(R)"             "IPS(R)"            "GOR(R)"           
#>  [28] "RUB(R)"            "mALT(L)"           "g1(R)"            
#>  [31] "CRE(R)"            "EB"                "FLA(R)"           
#>  [34] "SIP(R)"            "EPA(L)"            "PRW"              
#>  [37] "LAL(-GA)(R)"       "PB(L7)"            "VES(R)"           
#>  [40] "FBl7"              "FBl9"              "AOT(R)"           
#>  [43] "NO1(L)"            "NO(R)"             "SAD(-AMMC)"       
#>  [46] "FBl8"              "INP"               "RUB(L)"           
#>  [49] "VES(L)"            "NO2(L)"            "GOR(L)"           
#>  [52] "g5(R)"             "hemibrain"         "EBr6"             
#>  [55] "gL(L)"             "BU(R)"             "GNG"              
#>  [58] "PENP"              "PB(R8)"            "EBr3d"            
#>  [61] "PB(L6)"            "ICL(R)"            "b'1(R)"           
#>  [64] "SPS(L)"            "FBl1"              "SMP(R)"           
#>  [67] "EBr1"              "g4(R)"             "NO"               
#>  [70] "a2(R)"             "PB(R1)"            "NO2(R)"           
#>  [73] "NO(L)"             "PB(R2)"            "SIP(L)"           
#>  [76] "FB"                "PB"                "ME(R)"            
#>  [79] "PB(R9)"            "bL(L)"             "SNP(L)"           
#>  [82] "SCL(L)"            "g3(R)"             "EBr2r4"           
#>  [85] "FBl3"              "SLP(R)"            "VLNP(R)"          
#>  [88] "LOP(R)"            "SMP(L)"            "SAD"              
#>  [91] "WED(R)"            "PB(L8)"            "ATL(R)"           
#>  [94] "OL(R)"             "MB(R)"             "PB(R6)"           
#>  [97] "NO3(R)"            "b2(R)"             "SCL(R)"           
#> [100] "SNP(R)"            "aL(L)"             "CX"               
#> [103] "b'L(R)"            "g2(R)"             "PB(L1)"           
#> [106] "VMNP"              "EBr3pw"            "MB(+ACA)(R)"      
#> [109] "GF(R)"             "BU(L)"             "gL(R)"            
#> [112] "AOTU(R)"           "PB(R7)"            "AVLP(R)"          
#> [115] "CRE(-ROB,-RUB)(R)" "CA(L)"             "ROB(R)"           
#> [118] "AB(R)"             "LX(L)"             "LAL(R)"           
#> [121] "IB"                "PB(L3)"            "SPS(R)"           
#> [124] "aL(R)"             "AL(L)"             "EBr3am"           
#> [127] "CRE(L)"            "FBl6"              "mALT(R)"          
#> [130] "PB(L5)"            "AMMC"              "GA(R)"            
#> [133] "EPA(R)"            "PVLP(R)"           "CAN(R)"           
#> [136] "a3(R)"             "ATL(L)"            "a'2(R)"           
#> [139] "PLP(R)"            "PB(R5)"            "a1(R)"            
#> [142] "vACA(R)"           "LAL(L)"            "NO1(R)"           
#> [145] "LO(R)"             "b1(R)"             "a'L(L)"           
#> [148] "POC"               "PB(L2)"            "PED(R)"           
#> [151] "PB(L4)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "b'2(R)"            "CRE(L)"            "RUB(L)"           
#>   [4] "AL-VC4(R)"         "GOR(L)"            "WED(R)"           
#>   [7] "BU(R)"             "FBl1"              "AL-VC3l(R)"       
#>  [10] "MB(L)"             "EBr1"              "EB"               
#>  [13] "FB-column3"        "IB"                "SCL(L)"           
#>  [16] "gL(L)"             "AL-D(R)"           "CA(L)"            
#>  [19] "AOTU(R)"           "AL-DM5(L)"         "NO3(L)"           
#>  [22] "AL-VA2(R)"         "AL-DM1(L)"         "OL(R)"            
#>  [25] "AB(L)"             "AL-VM1(R)"         "AL-DA4l(R)"       
#>  [28] "AL-DL2d(R)"        "SMP(R)"            "b'1(R)"           
#>  [31] "SMP(L)"            "CX"                "PB(L8)"           
#>  [34] "LX(L)"             "AL-VP1l(R)"        "AL-DL5(L)"        
#>  [37] "a3(R)"             "AL-DM1(R)"         "a'2(R)"           
#>  [40] "NO"                "LO(R)"             "AL-VL2p(R)"       
#>  [43] "AL-VM4(R)"         "AL-VM7d(R)"        "a1(R)"            
#>  [46] "PB(R8)"            "AL-DM2(L)"         "CRE(-ROB,-RUB)(R)"
#>  [49] "PRW"               "PB(L6)"            "EBr3pw"           
#>  [52] "AL-DC2(R)"         "AL-VA3(R)"         "AL-VP5(R)"        
#>  [55] "PB(L1)"            "AL-DM6(L)"         "PB(L7)"           
#>  [58] "LH(R)"             "LAL(R)"            "AL-VC2(R)"        
#>  [61] "FBl6"              "AL-VC5(R)"         "POC"              
#>  [64] "CRE(-RUB)(L)"      "VES(R)"            "CA(R)"            
#>  [67] "AL-DL1(R)"         "ATL(R)"            "NO(L)"            
#>  [70] "AL-DC4(R)"         "g3(R)"             "lACA(R)"          
#>  [73] "mALT(R)"           "b2(R)"             "LOP(R)"           
#>  [76] "a'3(R)"            "AMMC"              "AL-DA2(L)"        
#>  [79] "aL(L)"             "SLP(R)"            "AL-VA5(R)"        
#>  [82] "SIP(L)"            "AL-DL4(R)"         "AL-DM4(R)"        
#>  [85] "AL-VL1(R)"         "GA(R)"             "PB"               
#>  [88] "LX(R)"             "EBr3d"             "INP"              
#>  [91] "AL-DC3(R)"         "AL-V(R)"           "a'1(R)"           
#>  [94] "AL-DL3(R)"         "AL-DP1m(R)"        "AL-VL2a(R)"       
#>  [97] "VLNP(R)"           "mALT(L)"           "GC"               
#> [100] "NO1(L)"            "BU(L)"             "AL(L)"            
#> [103] "bL(R)"             "EBr3am"            "g1(R)"            
#> [106] "EBr5"              "FLA(R)"            "AL-DM3(L)"        
#> [109] "AL-DP1m(L)"        "NO(R)"             "AL-VM7v(L)"       
#> [112] "AL-VA1d(R)"        "ATL(L)"            "AL-VM5d(R)"       
#> [115] "AL-VP2(R)"         "a2(R)"             "hemibrain"        
#> [118] "EBr6"              "AL-VP1d(R)"        "vACA(R)"          
#> [121] "AL-VA7l(R)"        "PB(L9)"            "AL-DA3(R)"        
#> [124] "aL(R)"             "gL(R)"             "AL-VM7v(R)"       
#> [127] "EBr2r4"            "ICL(L)"            "FBl4"             
#> [130] "MB(R)"             "PB(L5)"            "ICL(R)"           
#> [133] "PB(R5)"            "bL(L)"             "IPS(R)"           
#> [136] "AL(R)"             "NO2(L)"            "SPS(L)"           
#> [139] "PB(R1)"            "AL-VC3m(R)"        "PB(L3)"           
#> [142] "g2(R)"             "a'L(L)"            "AL-DC1(R)"        
#> [145] "AL-VC1(R)"         "RUB(R)"            "PB(R6)"           
#> [148] "GF(R)"             "AL-VM7d(L)"        "AL-VP4(R)"        
#> [151] "PLP(R)"            "FBl5"              "GNG"              
#> [154] "PB(L4)"            "AL-DA4m(R)"        "AL-DM3(R)"        
#> [157] "AL-VP3(R)"         "AL-VA1v(R)"        "PB(R7)"           
#> [160] "LAL(-GA)(R)"       "EPA(R)"            "ROB(R)"           
#> [163] "AME(R)"            "EPA(L)"            "GOR(R)"           
#> [166] "FBl3"              "AL-VM3(R)"         "SAD(-AMMC)"       
#> [169] "AL-DL2v(R)"        "AL-VA6(R)"         "AL-DA1(R)"        
#> [172] "SIP(R)"            "g5(R)"             "CRE(R)"           
#> [175] "SNP(R)"            "AL-DC1(L)"         "CAN(R)"           
#> [178] "PB(R4)"            "AL-DL4(L)"         "FBl2"             
#> [181] "PB(R2)"            "b'L(L)"            "AL-VA7m(R)"       
#> [184] "AL-VM2(R)"         "PENP"              "NO1(R)"           
#> [187] "a'L(R)"            "AL-DM2(R)"         "AB(R)"            
#> [190] "ME(R)"             "AL-DA4m(L)"        "AL-DC4(L)"        
#> [193] "PB(R9)"            "MB(+ACA)(R)"       "PB(R3)"           
#> [196] "b1(R)"             "LAL(L)"            "VES(L)"           
#> [199] "AL-VP1m(R)"        "NO3(R)"            "VMNP"             
#> [202] "FBl8"              "AVLP(R)"           "AL-DL5(R)"        
#> [205] "AL-DM4(L)"         "AL-VA6(L)"         "FB"               
#> [208] "AL-DA3(L)"         "FBl9"              "AL-DM6(R)"        
#> [211] "AL-VA4(R)"         "PB(L2)"            "b'L(R)"           
#> [214] "FBl7"              "AL-VM5v(R)"        "NO2(R)"           
#> [217] "SCL(R)"            "AL-D(L)"           "PVLP(R)"          
#> [220] "AL-DP1l(R)"        "SAD"               "SPS(R)"           
#> [223] "PED(R)"            "dACA(R)"           "g4(R)"            
#> [226] "AL-DA2(R)"         "AL-DM5(R)"         "SNP(L)"           
#> [229] "AOT(R)"            "AL-DC2(L)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-DA4m(L)"        "AL-V(R)"           "AL-DL1(R)"        
#>   [4] "AOTU(R)"           "SPS(R)"            "WED(R)"           
#>   [7] "RUB(L)"            "AL-VM5v(R)"        "SNP(R)"           
#>  [10] "AL-DM3(R)"         "PB(L8)"            "SPS(L)"           
#>  [13] "g5(R)"             "PB(R4)"            "GOR(L)"           
#>  [16] "b2(R)"             "AL-VM4(R)"         "PB(L2)"           
#>  [19] "EBr1"              "BU(R)"             "AL(L)"            
#>  [22] "FBl3"              "AL-VP1m(R)"        "PB(L1)"           
#>  [25] "AL-DC4(L)"         "FBl6"              "AL-VM3(R)"        
#>  [28] "g1(R)"             "PB(L6)"            "AME(R)"           
#>  [31] "AL-VP2(R)"         "AL-DM3(L)"         "AL-DM5(R)"        
#>  [34] "hemibrain"         "SCL(R)"            "NO1(R)"           
#>  [37] "VMNP"              "gL(L)"             "AL-DA1(R)"        
#>  [40] "AL-DL4(R)"         "mALT(R)"           "SNP(L)"           
#>  [43] "FBl7"              "AL-VM7v(R)"        "AL-DA3(L)"        
#>  [46] "NO(R)"             "FBl2"              "VLNP(R)"          
#>  [49] "ICL(R)"            "SLP(R)"            "AL-DL2v(R)"       
#>  [52] "AL-VM2(R)"         "AL-VP4(R)"         "AL-DC1(R)"        
#>  [55] "BU(L)"             "AL-DC1(L)"         "NO(L)"            
#>  [58] "OL(R)"             "AL-DA4m(R)"        "AL-VL2p(R)"       
#>  [61] "AL-VP3(R)"         "SIP(L)"            "a'L(R)"           
#>  [64] "AL-DM6(R)"         "EBr3am"            "b'2(R)"           
#>  [67] "AL-DL5(L)"         "AVLP(R)"           "EB"               
#>  [70] "AB(R)"             "AL-VP5(R)"         "NO3(R)"           
#>  [73] "PB(R5)"            "EBr2r4"            "ROB(R)"           
#>  [76] "NO3(L)"            "AL-VA5(R)"         "AL-VM7v(L)"       
#>  [79] "FBl8"              "INP"               "AL(R)"            
#>  [82] "AL-VA7l(R)"        "aL(L)"             "EPA(L)"           
#>  [85] "AL-VM5d(R)"        "vACA(R)"           "EBr3pw"           
#>  [88] "SAD(-AMMC)"        "AL-VA7m(R)"        "aL(R)"            
#>  [91] "FB"                "b'L(R)"            "PB(L9)"           
#>  [94] "GF(R)"             "PB(R8)"            "AL-DM5(L)"        
#>  [97] "FBl9"              "AL-VA1d(R)"        "PB(R6)"           
#> [100] "g4(R)"             "MB(+ACA)(R)"       "CA(R)"            
#> [103] "MB(R)"             "CRE(R)"            "AL-VM7d(L)"       
#> [106] "AL-VC1(R)"         "AL-VC3m(R)"        "AL-VA6(R)"        
#> [109] "IB"                "PLP(R)"            "AL-DP1m(L)"       
#> [112] "gL(R)"             "AL-DL2d(R)"        "AL-DM2(R)"        
#> [115] "AL-VC3l(R)"        "LOP(R)"            "PB(R1)"           
#> [118] "IPS(R)"            "a'L(L)"            "CA(L)"            
#> [121] "PENP"              "FB-column3"        "AL-DM1(R)"        
#> [124] "AB(L)"             "SCL(L)"            "PB"               
#> [127] "VES(R)"            "LH(R)"             "PB(L4)"           
#> [130] "AL-VL1(R)"         "SAD"               "PB(R7)"           
#> [133] "AL-DM6(L)"         "AL-VL2a(R)"        "mALT(L)"          
#> [136] "ATL(L)"            "a'2(R)"            "PB(R3)"           
#> [139] "b1(R)"             "AL-DL4(L)"         "CRE(-RUB)(L)"     
#> [142] "AL-DA4l(R)"        "AMMC"              "AL-DM4(L)"        
#> [145] "PB(L7)"            "PVLP(R)"           "LAL(R)"           
#> [148] "AL-DP1l(R)"        "NO2(L)"            "SMP(L)"           
#> [151] "GC"                "ATL(R)"            "bL(R)"            
#> [154] "g3(R)"             "MB(L)"             "FLA(R)"           
#> [157] "AL-VA3(R)"         "GA(R)"             "PRW"              
#> [160] "AL-DA2(L)"         "GNG"               "AL-DA3(R)"        
#> [163] "NO"                "LX(L)"             "CAN(R)"           
#> [166] "a1(R)"             "EBr5"              "EPA(R)"           
#> [169] "AL-VA6(L)"         "PB(R9)"            "LO(R)"            
#> [172] "AL-VP1l(R)"        "NO1(L)"            "ICL(L)"           
#> [175] "AL-VM7d(R)"        "lACA(R)"           "SMP(R)"           
#> [178] "EBr6"              "PB(R2)"            "AL-DC3(R)"        
#> [181] "AL-VA4(R)"         "dACA(R)"           "SIP(R)"           
#> [184] "bL(L)"             "AL-DA2(R)"         "AL-VA1v(R)"       
#> [187] "LX(R)"             "ME(R)"             "FBl5"             
#> [190] "NO2(R)"            "a3(R)"             "AL-VA2(R)"        
#> [193] "PED(R)"            "LAL(-GA)(R)"       "AL-D(L)"          
#> [196] "AL-DC4(R)"         "AL-DM4(R)"         "CRE(L)"           
#> [199] "EBr3d"             "AL-DM2(L)"         "FBl4"             
#> [202] "AL-VC2(R)"         "AL-VP1d(R)"        "RUB(R)"           
#> [205] "b'L(L)"            "VES(L)"            "AOT(R)"           
#> [208] "LAL(L)"            "AL-DC2(R)"         "AL-DL5(R)"        
#> [211] "AL-VC4(R)"         "POC"               "a'3(R)"           
#> [214] "g2(R)"             "b'1(R)"            "PB(L5)"           
#> [217] "AL-DM1(L)"         "AL-DP1m(R)"        "AL-VC5(R)"        
#> [220] "CX"                "a2(R)"             "AL-DC2(L)"        
#> [223] "a'1(R)"            "FBl1"              "AL-D(R)"          
#> [226] "GOR(R)"            "CRE(-ROB,-RUB)(R)" "PB(L3)"           
#> [229] "AL-DL3(R)"         "AL-VM1(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                          
#> [1,] "POST" "/api/npexplorer/neuronmeta"                  
#> [2,] "POST" "/api/raw/cypher/transaction"                 
#> [3,] "POST" "/api/raw/cypher/transaction/:id/cypher"      
#> [4,] "POST" "/api/v:ver/raw/cypher/transaction/:id/cypher"
#> [5,] "GET"  "/api/v:ver/npexplorer/rankedtable"           
#> [6,] "HEAD" "/api/*"                                      
#>      name                                                                                      
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMeta-fm"
#> [2,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                  
#> [3,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                  
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                      
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                      
#> [6,] "github.com/labstack/echo.glob..func1"                                                    

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }