Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "FBl8"              "LOP(R)"            "a'1(R)"           
#>   [4] "NO3(L)"            "LAL(-GA)(R)"       "WED(R)"           
#>   [7] "NO1(R)"            "SNP(R)"            "EB"               
#>  [10] "AME(R)"            "PB(R6)"            "lACA(R)"          
#>  [13] "a3(R)"             "MB(R)"             "SNP(L)"           
#>  [16] "b'L(L)"            "CX"                "NO3(R)"           
#>  [19] "ROB(R)"            "FBl6"              "bL(R)"            
#>  [22] "SCL(L)"            "EBr1"              "gL(L)"            
#>  [25] "FBl3"              "PRW"               "LX(R)"            
#>  [28] "SPS(L)"            "GF(R)"             "g3(R)"            
#>  [31] "CAN(R)"            "RUB(R)"            "BU(L)"            
#>  [34] "PVLP(R)"           "EBr3pw"            "AB(R)"            
#>  [37] "AOT(R)"            "PB"                "AL(L)"            
#>  [40] "PB(L7)"            "AB(L)"             "dACA(R)"          
#>  [43] "FBl1"              "ATL(R)"            "b'1(R)"           
#>  [46] "PB(L5)"            "CA(L)"             "SIP(L)"           
#>  [49] "CRE(R)"            "BU(R)"             "PB(L6)"           
#>  [52] "AVLP(R)"           "PB(L4)"            "SLP(R)"           
#>  [55] "vACA(R)"           "MB(L)"             "PB(L8)"           
#>  [58] "PB(L9)"            "ICL(L)"            "FBl7"             
#>  [61] "EPA(L)"            "EBr2r4"            "EPA(R)"           
#>  [64] "VES(R)"            "CA(R)"             "ATL(L)"           
#>  [67] "a1(R)"             "ME(R)"             "NO(R)"            
#>  [70] "NO2(L)"            "EBr3d"             "VLNP(R)"          
#>  [73] "PB(R8)"            "PB(L3)"            "b2(R)"            
#>  [76] "NO(L)"             "PB(R5)"            "g4(R)"            
#>  [79] "gL(R)"             "IPS(R)"            "EBr3am"           
#>  [82] "FBl5"              "ICL(R)"            "NO"               
#>  [85] "PB(R2)"            "INP"               "VES(L)"           
#>  [88] "PB(L2)"            "SAD"               "b'L(R)"           
#>  [91] "aL(R)"             "LX(L)"             "a'L(L)"           
#>  [94] "AOTU(R)"           "GC"                "OL(R)"            
#>  [97] "a'L(R)"            "g5(R)"             "a'2(R)"           
#> [100] "GA(R)"             "SPS(R)"            "PB(R9)"           
#> [103] "PB(R3)"            "LAL(L)"            "PB(R7)"           
#> [106] "GNG"               "POC"               "mALT(L)"          
#> [109] "AMMC"              "hemibrain"         "FB"               
#> [112] "SAD(-AMMC)"        "IB"                "bL(L)"            
#> [115] "MB(+ACA)(R)"       "PB(R4)"            "VMNP"             
#> [118] "SIP(R)"            "PB(R1)"            "CRE(L)"           
#> [121] "LAL(R)"            "PENP"              "b'2(R)"           
#> [124] "SMP(R)"            "PLP(R)"            "GOR(R)"           
#> [127] "NO2(R)"            "RUB(L)"            "FBl4"             
#> [130] "b1(R)"             "GOR(L)"            "EBr6"             
#> [133] "aL(L)"             "a2(R)"             "PED(R)"           
#> [136] "FBl9"              "LH(R)"             "AL(R)"            
#> [139] "EBr5"              "FBl2"              "PB(L1)"           
#> [142] "LO(R)"             "a'3(R)"            "mALT(R)"          
#> [145] "g1(R)"             "g2(R)"             "SCL(R)"           
#> [148] "FLA(R)"            "CRE(-ROB,-RUB)(R)" "SMP(L)"           
#> [151] "NO1(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                          
#> [1,] "GET"  "/api/v:ver/dbmeta/version"                   
#> [2,] "GET"  "/api/v:ver/custom/custom"                    
#> [3,] "GET"  "/api/npexplorer/findneurons"                 
#> [4,] "GET"  "/api/npexplorer/autapses"                    
#> [5,] "GET"  "/api/v:ver/npexplorer/autapses"              
#> [6,] "POST" "/api/v:ver/raw/cypher/transaction/:id/cypher"
#>      name                                                                                       
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getFindNeurons-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"   
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }