Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "mALT(R)"           "WED(R)"            "g5(R)"            
#>   [4] "GNG"               "RUB(R)"            "LAL(L)"           
#>   [7] "PB(R7)"            "CRE(L)"            "AB(L)"            
#>  [10] "FB"                "AL(L)"             "PB(R9)"           
#>  [13] "GOR(R)"            "SCL(L)"            "PENP"             
#>  [16] "GF(R)"             "lACA(R)"           "PVLP(R)"          
#>  [19] "IPS(R)"            "a'L(L)"            "SLP(R)"           
#>  [22] "SPS(R)"            "LAL(-GA)(R)"       "BU(L)"            
#>  [25] "PB(R8)"            "EPA(R)"            "AME(R)"           
#>  [28] "b'L(L)"            "PRW"               "SPS(L)"           
#>  [31] "MB(R)"             "PB(R5)"            "EBr2r4"           
#>  [34] "POC"               "PB"                "VMNP"             
#>  [37] "PB(R3)"            "EBr3am"            "PB(L4)"           
#>  [40] "SMP(L)"            "PB(L3)"            "CRE(R)"           
#>  [43] "b'L(R)"            "ME(R)"             "a'1(R)"           
#>  [46] "FBl3"              "gL(R)"             "SNP(L)"           
#>  [49] "AOTU(R)"           "PB(L8)"            "LX(R)"            
#>  [52] "gL(L)"             "NO3(L)"            "FBl1"             
#>  [55] "OL(R)"             "MB(L)"             "VES(R)"           
#>  [58] "CRE(-ROB,-RUB)(R)" "mALT(L)"           "aL(R)"            
#>  [61] "a'2(R)"            "ATL(R)"            "LX(L)"            
#>  [64] "VES(L)"            "CA(L)"             "PED(R)"           
#>  [67] "EBr5"              "g4(R)"             "PB(R2)"           
#>  [70] "bL(R)"             "SMP(R)"            "vACA(R)"          
#>  [73] "FBl8"              "VLNP(R)"           "PB(L5)"           
#>  [76] "hemibrain"         "GC"                "PB(L1)"           
#>  [79] "a'L(R)"            "AVLP(R)"           "LAL(R)"           
#>  [82] "CX"                "a2(R)"             "g1(R)"            
#>  [85] "ROB(R)"            "IB"                "a1(R)"            
#>  [88] "AOT(R)"            "NO2(L)"            "SCL(R)"           
#>  [91] "CAN(R)"            "b'1(R)"            "PB(R6)"           
#>  [94] "LO(R)"             "FBl7"              "AMMC"             
#>  [97] "a'3(R)"            "SIP(L)"            "b2(R)"            
#> [100] "EB"                "CA(R)"             "INP"              
#> [103] "a3(R)"             "NO(L)"             "ATL(L)"           
#> [106] "GOR(L)"            "NO"                "PB(L6)"           
#> [109] "bL(L)"             "MB(+ACA)(R)"       "RUB(L)"           
#> [112] "SAD"               "dACA(R)"           "BU(R)"            
#> [115] "SIP(R)"            "PB(L7)"            "FBl5"             
#> [118] "NO2(R)"            "PB(R4)"            "NO1(R)"           
#> [121] "PB(R1)"            "NO(R)"             "AL(R)"            
#> [124] "PB(L2)"            "EBr6"              "EBr3pw"           
#> [127] "FBl4"              "FBl9"              "GA(R)"            
#> [130] "EBr1"              "EBr3d"             "LH(R)"            
#> [133] "g3(R)"             "PB(L9)"            "FLA(R)"           
#> [136] "PLP(R)"            "ICL(R)"            "LOP(R)"           
#> [139] "b'2(R)"            "NO3(R)"            "g2(R)"            
#> [142] "FBl6"              "b1(R)"             "AB(R)"            
#> [145] "aL(L)"             "FBl2"              "NO1(L)"           
#> [148] "SNP(R)"            "ICL(L)"            "SAD(-AMMC)"       
#> [151] "EPA(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-VM7v(R)"        "ATL(R)"            "AL-VM4(R)"        
#>   [4] "AL-DA2(L)"         "AL-DA4l(R)"        "ICL(L)"           
#>   [7] "AB(R)"             "AL-DA3(R)"         "AL-DM6(R)"        
#>  [10] "AL-VA7m(R)"        "AL-VL2a(R)"        "bL(L)"            
#>  [13] "SAD(-AMMC)"        "a3(R)"             "CRE(L)"           
#>  [16] "AL-DL5(L)"         "MB(+ACA)(R)"       "VES(L)"           
#>  [19] "AL-DL5(R)"         "AL-VM3(R)"         "NO3(R)"           
#>  [22] "SCL(R)"            "AL-DC1(L)"         "FB-column3"       
#>  [25] "FBl3"              "OL(R)"             "AL-DP1m(R)"       
#>  [28] "AL-VC3l(R)"        "PB(R6)"            "AL-VP3(R)"        
#>  [31] "AL-DM1(L)"         "FBl5"              "AL-DC4(R)"        
#>  [34] "AL-VM5v(R)"        "EBr6"              "BU(L)"            
#>  [37] "LO(R)"             "RUB(L)"            "AL-VA1v(R)"       
#>  [40] "AL-VP5(R)"         "LAL(-GA)(R)"       "AL-VP1l(R)"       
#>  [43] "VES(R)"            "AL-VC3m(R)"        "NO1(R)"           
#>  [46] "hemibrain"         "mALT(R)"           "AL-DM5(L)"        
#>  [49] "PB(R3)"            "INP"               "AL-DC3(R)"        
#>  [52] "AL-DM5(R)"         "GA(R)"             "AL-DL1(R)"        
#>  [55] "AL-D(R)"           "AL-DA4m(R)"        "PB"               
#>  [58] "AL-VA4(R)"         "SPS(L)"            "EBr2r4"           
#>  [61] "gL(L)"             "AL-DL4(L)"         "ICL(R)"           
#>  [64] "CA(L)"             "g1(R)"             "CA(R)"            
#>  [67] "g5(R)"             "a'2(R)"            "EPA(R)"           
#>  [70] "AL(L)"             "b'L(L)"            "mALT(L)"          
#>  [73] "a2(R)"             "AL-VA3(R)"         "AL-VC1(R)"        
#>  [76] "SIP(L)"            "PB(L1)"            "AL-DA4m(L)"       
#>  [79] "AL-DP1m(L)"        "LX(L)"             "GOR(R)"           
#>  [82] "EBr3am"            "GOR(L)"            "AB(L)"            
#>  [85] "AL-VP2(R)"         "g3(R)"             "a1(R)"            
#>  [88] "b'1(R)"            "FBl8"              "AME(R)"           
#>  [91] "AL-VP4(R)"         "gL(R)"             "a'L(L)"           
#>  [94] "IB"                "CRE(R)"            "PB(L9)"           
#>  [97] "PVLP(R)"           "SPS(R)"            "AL-DA1(R)"        
#> [100] "AL-DM1(R)"         "PB(L5)"            "PB(R9)"           
#> [103] "PB(R2)"            "GC"                "PLP(R)"           
#> [106] "LH(R)"             "AL-VC4(R)"         "AL-VP1m(R)"       
#> [109] "SAD"               "FBl6"              "AL-VA1d(R)"       
#> [112] "AL-VC2(R)"         "AL-VM7d(R)"        "AL-DC2(L)"        
#> [115] "POC"               "RUB(R)"            "AOTU(R)"          
#> [118] "LX(R)"             "AL-DM3(L)"         "SIP(R)"           
#> [121] "FBl2"              "GNG"               "g2(R)"            
#> [124] "AL-DP1l(R)"        "VLNP(R)"           "AOT(R)"           
#> [127] "PB(R4)"            "ROB(R)"            "FBl4"             
#> [130] "a'L(R)"            "LAL(R)"            "NO2(L)"           
#> [133] "WED(R)"            "MB(R)"             "AL-DM2(R)"        
#> [136] "AL-VM1(R)"         "AL-VM7v(L)"        "NO3(L)"           
#> [139] "ME(R)"             "b2(R)"             "EBr3d"            
#> [142] "a'1(R)"            "AL-VA5(R)"         "PENP"             
#> [145] "SNP(L)"            "PB(L4)"            "EBr5"             
#> [148] "SCL(L)"            "NO1(L)"            "PB(L2)"           
#> [151] "AL-DC4(L)"         "AL-VA7l(R)"        "AL-VM2(R)"        
#> [154] "PB(R5)"            "AL-DL2v(R)"        "AL-VL1(R)"        
#> [157] "CX"                "AL-DM2(L)"         "MB(L)"            
#> [160] "AL-DC1(R)"         "AL-VP1d(R)"        "PB(L3)"           
#> [163] "FB"                "EBr1"              "VMNP"             
#> [166] "PB(R1)"            "NO(L)"             "AL-DC2(R)"        
#> [169] "lACA(R)"           "AL-DM6(L)"         "LOP(R)"           
#> [172] "LAL(L)"            "AL-VA6(R)"         "SMP(R)"           
#> [175] "AL(R)"             "b1(R)"             "bL(R)"            
#> [178] "AL-DL3(R)"         "AL-DM4(R)"         "SNP(R)"           
#> [181] "PED(R)"            "dACA(R)"           "AL-VL2p(R)"       
#> [184] "PB(R7)"            "AL-DA3(L)"         "PB(L7)"           
#> [187] "EPA(L)"            "aL(R)"             "AL-DA2(R)"        
#> [190] "AL-VM5d(R)"        "ATL(L)"            "FBl1"             
#> [193] "PB(L6)"            "IPS(R)"            "AL-VA2(R)"        
#> [196] "CRE(-ROB,-RUB)(R)" "b'2(R)"            "BU(R)"            
#> [199] "SMP(L)"            "CRE(-RUB)(L)"      "aL(L)"            
#> [202] "AL-DM3(R)"         "FLA(R)"            "FBl9"             
#> [205] "AL-VA6(L)"         "CAN(R)"            "AL-VC5(R)"        
#> [208] "PB(R8)"            "AL-DL2d(R)"        "AMMC"             
#> [211] "NO2(R)"            "SLP(R)"            "g4(R)"            
#> [214] "AL-DM4(L)"         "AL-VM7d(L)"        "NO(R)"            
#> [217] "EBr3pw"            "AL-D(L)"           "PB(L8)"           
#> [220] "NO"                "FBl7"              "AL-V(R)"          
#> [223] "a'3(R)"            "PRW"               "AVLP(R)"          
#> [226] "AL-DL4(R)"         "vACA(R)"           "b'L(R)"           
#> [229] "EB"                "GF(R)"            
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method  path                                     
#> [1,] "PUT"   "/api"                                   
#> [2,] "TRACE" "/api"                                   
#> [3,] "POST"  "/api/custom/custom"                     
#> [4,] "GET"   "/api/npexplorer/roisinneuron"           
#> [5,] "GET"   "/api/roimeshes/mesh/:dataset/:roi"      
#> [6,] "GET"   "/api/v:ver/roimeshes/mesh/:dataset/:roi"
#>      name                                                                                        
#> [1,] "github.com/labstack/echo.glob..func1"                                                      
#> [2,] "github.com/labstack/echo.glob..func1"                                                      
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"             
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/roimeshes.masterAPI.getMesh-fm"            
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }