Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DM6(L)"         "aL(L)"             "AL-VL1(R)"        
#>   [4] "BU(L)"             "ME(R)"             "FBl5"             
#>   [7] "hemibrain"         "LO(R)"             "AL-VP5(R)"        
#>  [10] "AL-VA4(R)"         "AL-VM7v(R)"        "SCL(R)"           
#>  [13] "g4(R)"             "AB(R)"             "AL-DM5(R)"        
#>  [16] "ATL(L)"            "PB(R1)"            "ICL(R)"           
#>  [19] "AL-DM1(R)"         "AL-VP4(R)"         "CA(L)"            
#>  [22] "SPS(L)"            "AL-DA3(L)"         "GOR(R)"           
#>  [25] "SAD"               "PB(R8)"            "LAL(R)"           
#>  [28] "AL-DA1(R)"         "PB"                "AL-DL4(L)"        
#>  [31] "EBr3pw"            "a'L(L)"            "ROB(R)"           
#>  [34] "LX(L)"             "AL-VP3(R)"         "WED(R)"           
#>  [37] "LX(R)"             "AL-DA2(L)"         "AL-VP2(R)"        
#>  [40] "bL(R)"             "AL-DL2d(R)"        "AOT(R)"           
#>  [43] "AL-DL2v(R)"        "PED(R)"            "EB"               
#>  [46] "PB(R9)"            "GNG"               "AL-DA3(R)"        
#>  [49] "AB(L)"             "AL-VM2(R)"         "SCL(L)"           
#>  [52] "a'L(R)"            "CAN(R)"            "AL-VM1(R)"        
#>  [55] "AL-VC2(R)"         "AL-VM3(R)"         "GOR(L)"           
#>  [58] "a'1(R)"            "AL-D(R)"           "SMP(L)"           
#>  [61] "EBr5"              "AL-DP1m(L)"        "a3(R)"            
#>  [64] "a1(R)"             "PRW"               "MB(L)"            
#>  [67] "AL-VM7d(L)"        "a'3(R)"            "AMMC"             
#>  [70] "g1(R)"             "EBr6"              "b'1(R)"           
#>  [73] "PENP"              "EBr1"              "dACA(R)"          
#>  [76] "SMP(R)"            "PB(L8)"            "FBl1"             
#>  [79] "OL(R)"             "VES(L)"            "EBr2r4"           
#>  [82] "AL-VM7v(L)"        "LH(R)"             "AL-VC5(R)"        
#>  [85] "BU(R)"             "CRE(L)"            "AL-VA5(R)"        
#>  [88] "PB(R6)"            "SIP(L)"            "AL-VA1d(R)"       
#>  [91] "AL-VC1(R)"         "vACA(R)"           "CRE(-ROB,-RUB)(R)"
#>  [94] "VMNP"              "SNP(L)"            "AL-DM2(R)"        
#>  [97] "EPA(L)"            "PB(R4)"            "MB(R)"            
#> [100] "AL-DA4m(L)"        "ICL(L)"            "AME(R)"           
#> [103] "PB(L4)"            "AL-VM5v(R)"        "POC"              
#> [106] "a'2(R)"            "CRE(-RUB)(L)"      "AL-DA4l(R)"       
#> [109] "AL-DA4m(R)"        "AL-DP1m(R)"        "AL-VL2p(R)"       
#> [112] "FB"                "AL-DC4(L)"         "FBl4"             
#> [115] "SIP(R)"            "CA(R)"             "AL-VA1v(R)"       
#> [118] "LOP(R)"            "MB(+ACA)(R)"       "ATL(R)"           
#> [121] "AL-DM3(R)"         "AL-DP1l(R)"        "EBr3am"           
#> [124] "AL-DC1(L)"         "AL-DC2(R)"         "AL-DM6(R)"        
#> [127] "RUB(L)"            "AL-VA3(R)"         "PB(R5)"           
#> [130] "AL-VP1d(R)"        "LAL(-GA)(R)"       "NO(L)"            
#> [133] "AL-DA2(R)"         "NO3(L)"            "AL-VC3m(R)"       
#> [136] "EBr3d"             "gL(L)"             "AL(L)"            
#> [139] "IB112"             "PB(R2)"            "AL-VP1m(R)"       
#> [142] "CRE(R)"            "AL-DC4(R)"         "PB(L2)"           
#> [145] "AL-DL5(L)"         "AL-DM4(L)"         "bL(L)"            
#> [148] "AL-V(R)"           "AL-DM4(R)"         "AL-VA7l(R)"       
#> [151] "PB(L3)"            "AL-DM3(L)"         "AL-DM5(L)"        
#> [154] "FBl9"              "AL-VM7d(R)"        "VLNP(R)"          
#> [157] "AL-DC2(L)"         "PB(L6)"            "gL(R)"            
#> [160] "AVLP(R)"           "AL-VA6(R)"         "AL-VC4(R)"        
#> [163] "GA(R)"             "GC"                "g5(R)"            
#> [166] "aL(R)"             "AL(R)"             "AL-VL2a(R)"       
#> [169] "b'2(R)"            "AOTU(R)"           "SPS(R)"           
#> [172] "NO(R)"             "AL-VM5d(R)"        "lACA(R)"          
#> [175] "PB(R7)"            "AL-VA6(L)"         "AL-DM2(L)"        
#> [178] "b'L(L)"            "AL-VA2(R)"         "AL-VP1l(R)"       
#> [181] "EPA(R)"            "SAD(-AMMC)"        "FBl3"             
#> [184] "NO1(L)"            "NO"                "AL-VM4(R)"        
#> [187] "FBl6"              "AL-DL4(R)"         "CX"               
#> [190] "LAL(L)"            "g3(R)"             "PLP(R)"           
#> [193] "AL-DL3(R)"         "b2(R)"             "PB(L1)"           
#> [196] "AL-DL1(R)"         "NO2(R)"            "AL-VC3l(R)"       
#> [199] "FLA(R)"            "AL-D(L)"           "FB-column3"       
#> [202] "IB"                "mALT(R)"           "g2(R)"            
#> [205] "SLP(R)"            "mALT(L)"           "INP"              
#> [208] "AL-DC1(R)"         "AL-DL5(R)"         "a2(R)"            
#> [211] "PB(L5)"            "FBl2"              "RUB(R)"           
#> [214] "NO1(R)"            "PB(L9)"            "VES(R)"           
#> [217] "FBl8"              "PB(R3)"            "IPS(R)"           
#> [220] "AL-DC3(R)"         "AL-VA7m(R)"        "GF(R)"            
#> [223] "NO3(R)"            "AL-DM1(L)"         "PVLP(R)"          
#> [226] "FBl7"              "b1(R)"             "NO2(L)"           
#> [229] "b'L(R)"            "SNP(R)"            "PB(L7)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "a'L(L)"            "CX"                "EPA(R)"           
#>   [4] "AL(L)"             "gL(R)"             "LO(R)"            
#>   [7] "aL(L)"             "a'3(R)"            "g1(R)"            
#>  [10] "FB"                "EBr5"              "EBr3pw"           
#>  [13] "FBl4"              "PB(L3)"            "CRE(R)"           
#>  [16] "SNP(R)"            "SPS(R)"            "BU(L)"            
#>  [19] "FBl5"              "OL(R)"             "MB(R)"            
#>  [22] "PED(R)"            "FBl7"              "POC"              
#>  [25] "RUB(R)"            "g3(R)"             "SIP(L)"           
#>  [28] "NO2(L)"            "b'1(R)"            "ME(R)"            
#>  [31] "PVLP(R)"           "AVLP(R)"           "GOR(R)"           
#>  [34] "b'L(L)"            "PB(L1)"            "LX(R)"            
#>  [37] "a'1(R)"            "LX(L)"             "bL(R)"            
#>  [40] "AOT(R)"            "LAL(-GA)(R)"       "GC"               
#>  [43] "EBr1"              "NO"                "VES(L)"           
#>  [46] "ATL(R)"            "NO2(R)"            "a1(R)"            
#>  [49] "LAL(L)"            "a3(R)"             "PB(L6)"           
#>  [52] "NO(R)"             "VLNP(R)"           "PB(L2)"           
#>  [55] "EBr3am"            "LAL(R)"            "SCL(L)"           
#>  [58] "LOP(R)"            "PENP"              "MB(L)"            
#>  [61] "SIP(R)"            "bL(L)"             "FBl1"             
#>  [64] "PRW"               "GOR(L)"            "EB"               
#>  [67] "CRE(L)"            "PB(L4)"            "SMP(R)"           
#>  [70] "CA(L)"             "NO1(L)"            "ROB(R)"           
#>  [73] "NO3(L)"            "SNP(L)"            "LH(R)"            
#>  [76] "CAN(R)"            "SLP(R)"            "ATL(L)"           
#>  [79] "SMP(L)"            "b'L(R)"            "PB(R8)"           
#>  [82] "EBr2r4"            "AME(R)"            "PB(R6)"           
#>  [85] "WED(R)"            "NO1(R)"            "a'L(R)"           
#>  [88] "mALT(R)"           "g5(R)"             "FBl8"             
#>  [91] "RUB(L)"            "GF(R)"             "EBr3d"            
#>  [94] "gL(L)"             "BU(R)"             "FBl2"             
#>  [97] "a2(R)"             "PLP(R)"            "g4(R)"            
#> [100] "SAD(-AMMC)"        "VES(R)"            "EPA(L)"           
#> [103] "AMMC"              "AOTU(R)"           "g2(R)"            
#> [106] "FBl9"              "CA(R)"             "AB(R)"            
#> [109] "FLA(R)"            "INP"               "GNG"              
#> [112] "VMNP"              "MB(+ACA)(R)"       "PB(L8)"           
#> [115] "PB(R3)"            "PB(R4)"            "AL(R)"            
#> [118] "vACA(R)"           "PB(R1)"            "PB(L7)"           
#> [121] "FBl3"              "PB(L5)"            "mALT(L)"          
#> [124] "EBr6"              "PB"                "PB(L9)"           
#> [127] "FBl6"              "AB(L)"             "lACA(R)"          
#> [130] "SPS(L)"            "PB(R9)"            "IPS(R)"           
#> [133] "CRE(-ROB,-RUB)(R)" "aL(R)"             "SCL(R)"           
#> [136] "a'2(R)"            "b'2(R)"            "GA(R)"            
#> [139] "hemibrain"         "SAD"               "PB(R7)"           
#> [142] "ICL(L)"            "NO3(R)"            "b2(R)"            
#> [145] "dACA(R)"           "PB(R2)"            "NO(L)"            
#> [148] "ICL(R)"            "b1(R)"             "PB(R5)"           
#> [151] "IB"               
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "b2(R)"             "EBr6"              "NO1(R)"           
#>   [4] "FBl1"              "AL-DA3(R)"         "AL-DM2(L)"        
#>   [7] "ICL(L)"            "CRE(-ROB,-RUB)(R)" "SPS(L)"           
#>  [10] "AL-VL2p(R)"        "AL-DL2d(R)"        "AL-VA1d(R)"       
#>  [13] "AL-VA7m(R)"        "MB(L)"             "EPA(L)"           
#>  [16] "AL-VA4(R)"         "LH(R)"             "AL-DC1(R)"        
#>  [19] "AL-VL1(R)"         "a2(R)"             "NO3(R)"           
#>  [22] "INP"               "LAL(R)"            "AL-DM4(R)"        
#>  [25] "FB-column3"        "AL-VA3(R)"         "AL-VA5(R)"        
#>  [28] "a'2(R)"            "VES(R)"            "SLP(R)"           
#>  [31] "AMMC"              "PENP"              "AL-DC3(R)"        
#>  [34] "LOP(R)"            "PB(L2)"            "b'2(R)"           
#>  [37] "PB(R7)"            "g5(R)"             "FBl9"             
#>  [40] "AL-VM2(R)"         "EBr5"              "dACA(R)"          
#>  [43] "AL-DM3(R)"         "AOT(R)"            "GOR(R)"           
#>  [46] "AL-DL5(R)"         "PB(R6)"            "AL-DC4(L)"        
#>  [49] "PB(R1)"            "LO(R)"             "NO2(R)"           
#>  [52] "NO1(L)"            "bL(L)"             "EBr3pw"           
#>  [55] "AL-VM7d(R)"        "a'L(R)"            "a'1(R)"           
#>  [58] "FBl3"              "AL-DM1(R)"         "GF(R)"            
#>  [61] "PB(L8)"            "PB(L4)"            "AB(L)"            
#>  [64] "AOTU(R)"           "AL-DM6(L)"         "AL-DP1m(L)"       
#>  [67] "FBl4"              "BU(R)"             "SNP(L)"           
#>  [70] "PED(R)"            "FLA(R)"            "VES(L)"           
#>  [73] "AL-DA2(R)"         "AL-VM5v(R)"        "EB"               
#>  [76] "AL-DA4m(L)"        "PB(L9)"            "ME(R)"            
#>  [79] "b'L(L)"            "PB(R5)"            "PRW"              
#>  [82] "lACA(R)"           "AL-DM5(L)"         "AL-VA2(R)"        
#>  [85] "AL-VM4(R)"         "AL-DM4(L)"         "SIP(R)"           
#>  [88] "NO3(L)"            "IPS(R)"            "ATL(L)"           
#>  [91] "SMP(L)"            "CRE(R)"            "g2(R)"            
#>  [94] "EBr2r4"            "AL-VM7d(L)"        "LAL(L)"           
#>  [97] "PB(R8)"            "ROB(R)"            "SAD"              
#> [100] "aL(R)"             "AL-VC3m(R)"        "PLP(R)"           
#> [103] "PB(L7)"            "hemibrain"         "SPS(R)"           
#> [106] "PB(R2)"            "AL-VL2a(R)"        "PB(R9)"           
#> [109] "FBl6"              "AL-DM5(R)"         "AL-VM1(R)"        
#> [112] "AL-VM7v(R)"        "AL-DA4l(R)"        "AL-DM6(R)"        
#> [115] "AL-VM5d(R)"        "AL-DL5(L)"         "AL-DM3(L)"        
#> [118] "AME(R)"            "LX(L)"             "OL(R)"            
#> [121] "AL-VA6(L)"         "AL-VC1(R)"         "AL-VP5(R)"        
#> [124] "aL(L)"             "PB(R4)"            "SCL(R)"           
#> [127] "BU(L)"             "AL-VA6(R)"         "SMP(R)"           
#> [130] "PB(L1)"            "gL(L)"             "AL-DL1(R)"        
#> [133] "AL-DL3(R)"         "AL-VC4(R)"         "AL-VP1l(R)"       
#> [136] "a1(R)"             "NO2(L)"            "AL-DA2(L)"        
#> [139] "FBl7"              "SCL(L)"            "GOR(L)"           
#> [142] "AL(L)"             "AL-DC1(L)"         "AL-DA4m(R)"       
#> [145] "AL-VP1m(R)"        "SAD(-AMMC)"        "AL-VC2(R)"        
#> [148] "AL-VC3l(R)"        "SNP(R)"            "AL-DA3(L)"        
#> [151] "PB(L6)"            "POC"               "vACA(R)"          
#> [154] "bL(R)"             "AL-DP1m(R)"        "CX"               
#> [157] "FB"                "AL-D(L)"           "FBl5"             
#> [160] "AL-D(R)"           "AL-VC5(R)"         "RUB(R)"           
#> [163] "g1(R)"             "LX(R)"             "NO(R)"            
#> [166] "FBl2"              "MB(R)"             "PB(L5)"           
#> [169] "g3(R)"             "LAL(-GA)(R)"       "AL-DM1(L)"        
#> [172] "AL(R)"             "VLNP(R)"           "GC"               
#> [175] "WED(R)"            "EBr1"              "AL-VP1d(R)"       
#> [178] "CRE(L)"            "CA(R)"             "ATL(R)"           
#> [181] "AL-VP3(R)"         "VMNP"              "PVLP(R)"          
#> [184] "gL(R)"             "PB(R3)"            "AL-DA1(R)"        
#> [187] "CAN(R)"            "AL-VM3(R)"         "a'3(R)"           
#> [190] "SIP(L)"            "b'L(R)"            "a'L(L)"           
#> [193] "AL-DL2v(R)"        "EPA(R)"            "NO(L)"            
#> [196] "AL-DC2(R)"         "AL-VA7l(R)"        "EBr3am"           
#> [199] "GA(R)"             "PB"                "MB(+ACA)(R)"      
#> [202] "AL-VA1v(R)"        "AL-DM2(R)"         "EBr3d"            
#> [205] "AL-DL4(L)"         "AL-VM7v(L)"        "FBl8"             
#> [208] "AB(R)"             "mALT(R)"           "g4(R)"            
#> [211] "AL-VP4(R)"         "AVLP(R)"           "GNG"              
#> [214] "a3(R)"             "AL-DC4(R)"         "AL-DL4(R)"        
#> [217] "b'1(R)"            "mALT(L)"           "PB(L3)"           
#> [220] "b1(R)"             "IB"                "AL-DC2(L)"        
#> [223] "CRE(-RUB)(L)"      "ICL(R)"            "AL-DP1l(R)"       
#> [226] "AL-VP2(R)"         "NO"                "RUB(L)"           
#> [229] "AL-V(R)"           "CA(L)"            
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "CA(L)"             "EBr1"              "AL-DM2(L)"        
#>   [4] "AL-DM4(L)"         "PB(L9)"            "FBl1"             
#>   [7] "AOT(R)"            "b'L(L)"            "AL-DP1m(R)"       
#>  [10] "a2(R)"             "AL-DM6(R)"         "AL-VM5v(R)"       
#>  [13] "AL-DC4(L)"         "AL-DM5(L)"         "AL-DP1m(L)"       
#>  [16] "AVLP(R)"           "LX(L)"             "AL-VP1m(R)"       
#>  [19] "GC"                "AL-VC3m(R)"        "MB(+ACA)(R)"      
#>  [22] "AL-DL3(R)"         "AL-VP2(R)"         "AL-VP1l(R)"       
#>  [25] "AL-VP1d(R)"        "b'L(R)"            "bL(L)"            
#>  [28] "AL-DM5(R)"         "VES(L)"            "b2(R)"            
#>  [31] "BU(L)"             "SNP(R)"            "OL(R)"            
#>  [34] "SCL(L)"            "BU(R)"             "b'1(R)"           
#>  [37] "LX(R)"             "FBl3"              "PB(L4)"           
#>  [40] "FB"                "SAD"               "AL-VP5(R)"        
#>  [43] "mALT(L)"           "AL-D(L)"           "NO"               
#>  [46] "gL(R)"             "AME(R)"            "EPA(L)"           
#>  [49] "AL-VM7d(L)"        "AL-VA6(R)"         "a1(R)"            
#>  [52] "AOTU(R)"           "SPS(L)"            "NO(L)"            
#>  [55] "AL-DA4l(R)"        "AL-VA4(R)"         "gL(L)"            
#>  [58] "FBl2"              "EBr2r4"            "AL-DL1(R)"        
#>  [61] "AL-VM1(R)"         "mALT(R)"           "PB(R6)"           
#>  [64] "a'L(R)"            "AL-DA3(L)"         "RUB(L)"           
#>  [67] "AMMC"              "NO1(R)"            "SIP(R)"           
#>  [70] "FBl5"              "FBl4"              "CAN(R)"           
#>  [73] "bL(R)"             "LAL(R)"            "POC"              
#>  [76] "g4(R)"             "PB(R9)"            "AB(L)"            
#>  [79] "AL-VM5d(R)"        "b'2(R)"            "GOR(L)"           
#>  [82] "AL-DA2(L)"         "AL-DL4(R)"         "AL-VA3(R)"        
#>  [85] "AL-VM7d(R)"        "RUB(R)"            "hemibrain"        
#>  [88] "PB(L8)"            "PLP(R)"            "AL-DL5(R)"        
#>  [91] "lACA(R)"           "AL-VA6(L)"         "PB(R1)"           
#>  [94] "AL-VM7v(R)"        "PRW"               "EBr3pw"           
#>  [97] "ATL(R)"            "NO2(R)"            "AL-VA7m(R)"       
#> [100] "GF(R)"             "a'2(R)"            "SAD(-AMMC)"       
#> [103] "AL-VL2a(R)"        "CX"                "AL-DM6(L)"        
#> [106] "AL-DA4m(R)"        "AL-VL1(R)"         "AL-VP4(R)"        
#> [109] "NO2(L)"            "g1(R)"             "PB(R8)"           
#> [112] "AL(L)"             "PB(L7)"            "VES(R)"           
#> [115] "INP"               "NO3(R)"            "EB"               
#> [118] "NO(R)"             "a'1(R)"            "AL(R)"            
#> [121] "AL-DP1l(R)"        "NO1(L)"            "AL-DM3(L)"        
#> [124] "LAL(-GA)(R)"       "AL-DL4(L)"         "SNP(L)"           
#> [127] "AL-VM4(R)"         "SPS(R)"            "EBr6"             
#> [130] "AL-VM7v(L)"        "g5(R)"             "AL-D(R)"          
#> [133] "FBl6"              "PB(R2)"            "SLP(R)"           
#> [136] "PB(R3)"            "MB(L)"             "PVLP(R)"          
#> [139] "AL-DM1(R)"         "WED(R)"            "dACA(R)"          
#> [142] "AL-DM4(R)"         "AL-VM2(R)"         "FBl8"             
#> [145] "aL(L)"             "FB-column3"        "IB"               
#> [148] "LAL(L)"            "SCL(R)"            "LH(R)"            
#> [151] "PB(R5)"            "GA(R)"             "aL(R)"            
#> [154] "AL-DM2(R)"         "AL-VC1(R)"         "AL-VM3(R)"        
#> [157] "CRE(-ROB,-RUB)(R)" "VMNP"              "CA(R)"            
#> [160] "AL-VA5(R)"         "CRE(R)"            "a3(R)"            
#> [163] "AL-VC5(R)"         "a'3(R)"            "PB(L3)"           
#> [166] "EBr5"              "PB(L1)"            "EBr3d"            
#> [169] "AL-DA2(R)"         "SMP(R)"            "AL-DM1(L)"        
#> [172] "b1(R)"             "AL-DC2(R)"         "AL-VA1v(R)"       
#> [175] "ROB(R)"            "AL-DC3(R)"         "AL-VA1d(R)"       
#> [178] "PB(L6)"            "EBr3am"            "FBl9"             
#> [181] "AL-DL2d(R)"        "AL-VC2(R)"         "AL-VC4(R)"        
#> [184] "AL-VL2p(R)"        "AB(R)"             "a'L(L)"           
#> [187] "LOP(R)"            "PED(R)"            "ICL(R)"           
#> [190] "PB(L5)"            "PB(R7)"            "AL-DA4m(L)"       
#> [193] "AL-DC2(L)"         "AL-DC4(R)"         "PB(L2)"           
#> [196] "AL-VC3l(R)"        "PB"                "CRE(L)"           
#> [199] "PB(R4)"            "AL-DM3(R)"         "AL-V(R)"          
#> [202] "VLNP(R)"           "AL-DA3(R)"         "vACA(R)"          
#> [205] "PENP"              "SMP(L)"            "EPA(R)"           
#> [208] "ICL(L)"            "FBl7"              "AL-DA1(R)"        
#> [211] "ME(R)"             "AL-DL5(L)"         "GOR(R)"           
#> [214] "SIP(L)"            "FLA(R)"            "LO(R)"            
#> [217] "AL-VA7l(R)"        "NO3(L)"            "AL-VA2(R)"        
#> [220] "AL-VP3(R)"         "AL-DC1(L)"         "GNG"              
#> [223] "MB(R)"             "AL-DL2v(R)"        "ATL(L)"           
#> [226] "g3(R)"             "g2(R)"             "CRE(-RUB)(L)"     
#> [229] "IPS(R)"            "AL-DC1(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`manc:v1.0`
#> $`manc:v1.0`$`last-mod`
#> [1] "2023-05-31 23:54:01"
#> 
#> $`manc:v1.0`$uuid
#> [1] "59b37970bc7a4341b9a3a965a0d6b402"
#> 
#> $`manc:v1.0`$ROIs
#>  [1] "AbN1(L)"          "AbN3(L)"          "CvN(L)"           "MesoAN(L)"       
#>  [5] "HTct(UTct-T3)(L)" "IntTct"           "GF(R)"            "MesoLN(L)"       
#>  [9] "mVAC(T1)(L)"      "mVAC(T3)(L)"      "mVAC(T3)(R)"      "NTct(UTct-T1)(R)"
#> [13] "AbN1(R)"          "LegNp(T2)(R)"     "AbN2(R)"          "CvN(R)"          
#> [17] "MetaLN(R)"        "PDMN(L)"          "LegNp(T3)(L)"     "ADMN(R)"         
#> [21] "ProAN(L)"         "VProN(L)"         "ProCN(R)"         "LegNp(T1)(R)"    
#> [25] "LegNp(T3)(R)"     "ProLN(L)"         "PrN(L)"           "LegNp(T2)(L)"    
#> [29] "mVAC(T2)(L)"      "LTct"             "DMetaN(R)"        "PDMN(R)"         
#> [33] "AbN2(L)"          "DProN(R)"         "ProAN(R)"         "LegNp(T1)(L)"    
#> [37] "mVAC(T1)(R)"      "AbN3(R)"          "MetaLN(L)"        "ProLN(R)"        
#> [41] "mVAC(T2)(R)"      "ANm"              "WTct(UTct-T2)(R)" "CV"              
#> [45] "HTct(UTct-T3)(R)" "GF(L)"            "MesoAN(R)"        "ADMN(L)"         
#> [49] "AbNT"             "DProN(L)"         "VProN(R)"         "Ov(L)"           
#> [53] "NTct(UTct-T1)(L)" "WTct(UTct-T2)(L)" "DMetaN(L)"        "ProCN(L)"        
#> [57] "AbN4(L)"          "PrN(R)"           "Ov(R)"            "AbN4(R)"         
#> [61] "MesoLN(R)"       
#> 
#> $`manc:v1.0`$superLevelROIs
#>  [1] "ADMN(L)"          "ADMN(R)"          "Ov(L)"            "Ov(R)"           
#>  [5] "ANm"              "AbN1(L)"          "AbN1(R)"          "AbN2(L)"         
#>  [9] "AbN2(R)"          "AbN3(L)"          "AbN3(R)"          "AbN4(L)"         
#> [13] "AbN4(R)"          "AbNT"             "CV"               "CvN(L)"          
#> [17] "CvN(R)"           "DMetaN(L)"        "DMetaN(R)"        "DProN(L)"        
#> [21] "DProN(R)"         "GF(L)"            "GF(R)"            "HTct(UTct-T3)(L)"
#> [25] "HTct(UTct-T3)(R)" "LegNp(T1)(L)"     "LegNp(T1)(R)"     "LegNp(T2)(L)"    
#> [29] "LegNp(T2)(R)"     "LegNp(T3)(L)"     "LegNp(T3)(R)"     "IntTct"          
#> [33] "LTct"             "MesoAN(L)"        "MesoAN(R)"        "MesoLN(L)"       
#> [37] "MesoLN(R)"        "MetaLN(L)"        "MetaLN(R)"        "NTct(UTct-T1)(L)"
#> [41] "NTct(UTct-T1)(R)" "PDMN(L)"          "PrN(L)"           "PrN(R)"          
#> [45] "ProCN(L)"         "ProCN(R)"         "PDMN(R)"          "ProAN(L)"        
#> [49] "ProAN(R)"         "ProLN(L)"         "ProLN(R)"         "VProN(L)"        
#> [53] "VProN(R)"         "WTct(UTct-T2)(L)" "WTct(UTct-T2)(R)" "mVAC(T1)(L)"     
#> [57] "mVAC(T1)(R)"      "mVAC(T2)(L)"      "mVAC(T2)(R)"      "mVAC(T3)(L)"     
#> [61] "mVAC(T3)(R)"     
#> 
#> $`manc:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/manc-connectome"
#> 
#> $`manc:v1.0`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                  
#> [1,] "GET"  "/api/npexplorer/distribution"        
#> [2,] "GET"  "/api/dbmeta/database"                
#> [3,] "GET"  "/api"                                
#> [4,] "GET"  "/api/serverinfo"                     
#> [5,] "GET"  "/api/npexplorer/nglayers*"           
#> [6,] "POST" "/api/skeletons/skeleton/:dataset/:id"
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatabase-fm"            
#> [3,] "github.com/labstack/echo/v4.glob..func1"                                                   
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [5,] "github.com/labstack/echo/v4.StaticDirectoryHandler.func1"                                  
#> [6,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }