Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "EBr1"              "PB(L1)"            "INP"              
#>   [4] "mALT(L)"           "MB(L)"             "WED(R)"           
#>   [7] "aL(R)"             "SCL(R)"            "NO(R)"            
#>  [10] "SAD(-AMMC)"        "EPA(L)"            "PB(L2)"           
#>  [13] "lACA(R)"           "PB(R8)"            "VLNP(R)"          
#>  [16] "AMMC"              "CX"                "PB(L3)"           
#>  [19] "FBl4"              "LX(L)"             "PB(R5)"           
#>  [22] "PB(L5)"            "EBr6"              "g2(R)"            
#>  [25] "a'2(R)"            "gL(R)"             "FBl8"             
#>  [28] "IPS(R)"            "IB"                "EBr3am"           
#>  [31] "NO1(L)"            "PVLP(R)"           "GNG"              
#>  [34] "POC"               "PB(L6)"            "mALT(R)"          
#>  [37] "LX(R)"             "FBl2"              "FBl5"             
#>  [40] "AME(R)"            "SAD"               "NO"               
#>  [43] "FBl7"              "PB(R3)"            "PLP(R)"           
#>  [46] "PB(R1)"            "AB(R)"             "bL(R)"            
#>  [49] "SLP(R)"            "GA(R)"             "SNP(R)"           
#>  [52] "AL(L)"             "ATL(R)"            "RUB(L)"           
#>  [55] "NO2(L)"            "PB"                "MB(+ACA)(R)"      
#>  [58] "FBl3"              "NO(L)"             "a'3(R)"           
#>  [61] "BU(L)"             "FBl6"              "LH(R)"            
#>  [64] "SNP(L)"            "a2(R)"             "PRW"              
#>  [67] "EBr5"              "bL(L)"             "GOR(R)"           
#>  [70] "SCL(L)"            "PB(R6)"            "VES(R)"           
#>  [73] "FBl9"              "AVLP(R)"           "PB(R4)"           
#>  [76] "a'L(L)"            "PB(L9)"            "PB(L7)"           
#>  [79] "b1(R)"             "aL(L)"             "ME(R)"            
#>  [82] "VES(L)"            "AOT(R)"            "RUB(R)"           
#>  [85] "EBr3d"             "FLA(R)"            "g4(R)"            
#>  [88] "ICL(R)"            "g1(R)"             "SIP(R)"           
#>  [91] "NO2(R)"            "b'1(R)"            "b'2(R)"           
#>  [94] "EPA(R)"            "CAN(R)"            "SMP(R)"           
#>  [97] "g3(R)"             "GOR(L)"            "LO(R)"            
#> [100] "LAL(R)"            "PB(L4)"            "vACA(R)"          
#> [103] "GF(R)"             "ICL(L)"            "g5(R)"            
#> [106] "MB(R)"             "EBr3pw"            "GC"               
#> [109] "FB"                "PB(L8)"            "NO1(R)"           
#> [112] "gL(L)"             "a3(R)"             "SIP(L)"           
#> [115] "EBr2r4"            "ROB(R)"            "ATL(L)"           
#> [118] "b'L(R)"            "NO3(R)"            "CRE(L)"           
#> [121] "PED(R)"            "LAL(-GA)(R)"       "PENP"             
#> [124] "SPS(L)"            "NO3(L)"            "LAL(L)"           
#> [127] "BU(R)"             "AL(R)"             "PB(R7)"           
#> [130] "AOTU(R)"           "hemibrain"         "dACA(R)"          
#> [133] "a'L(R)"            "a'1(R)"            "FBl1"             
#> [136] "PB(R2)"            "SMP(L)"            "CRE(-ROB,-RUB)(R)"
#> [139] "OL(R)"             "CA(L)"             "b'L(L)"           
#> [142] "CRE(R)"            "EB"                "CA(R)"            
#> [145] "LOP(R)"            "a1(R)"             "SPS(R)"           
#> [148] "VMNP"              "PB(R9)"            "AB(L)"            
#> [151] "b2(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                                  
#> [1,] "OPTIONS" "/api/*"                              
#> [2,] "POST"    "/api/v:ver/npexplorer/findneurons"   
#> [3,] "POST"    "/api/v:ver/npexplorer/completeness"  
#> [4,] "GET"     "/api/skeletons/skeleton/:dataset/:id"
#> [5,] "CONNECT" "/api/*"                              
#> [6,] "GET"     "/api/v:ver/cached/roiconnectivity"   
#>      name                                                                                
#> [1,] "github.com/labstack/echo.glob..func1"                                              
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/skeletons.masterAPI.getSkeleton-fm"
#> [5,] "github.com/labstack/echo.glob..func1"                                              
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }