Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "FB-column3"        "FBl4"              "PB(R6)"           
#>   [4] "FBl3"              "AL-VA3(R)"         "ROB(R)"           
#>   [7] "a3(R)"             "AL-DA2(R)"         "SMP(R)"           
#>  [10] "AL-DC2(L)"         "AL-DP1m(R)"        "mALT(L)"          
#>  [13] "AL-DM4(R)"         "AL-VA5(R)"         "POC"              
#>  [16] "a'3(R)"            "NO2(L)"            "AL-DA3(R)"        
#>  [19] "ICL(L)"            "FBl2"              "AL-DL5(R)"        
#>  [22] "AOTU(R)"           "LO(R)"             "LAL(R)"           
#>  [25] "VES(L)"            "EBr6"              "SLP(R)"           
#>  [28] "MB(R)"             "AL-DC1(R)"         "AL-DM2(R)"        
#>  [31] "lACA(R)"           "MB(L)"             "PB(L7)"           
#>  [34] "PB(L4)"            "AL-DA4m(R)"        "AL-VL2a(R)"       
#>  [37] "AL-VM7d(R)"        "mALT(R)"           "AL-DL2d(R)"       
#>  [40] "AL-DL5(L)"         "AL-VC2(R)"         "AL-VP1l(R)"       
#>  [43] "CA(L)"             "SNP(R)"            "AL-DC1(L)"        
#>  [46] "FBl5"              "AL-VC4(R)"         "AL-VM5d(R)"       
#>  [49] "g1(R)"             "dACA(R)"           "PB(R8)"           
#>  [52] "g4(R)"             "gL(R)"             "CRE(-RUB)(L)"     
#>  [55] "GOR(R)"            "AL-V(R)"           "SIP(L)"           
#>  [58] "FB"                "b'1(R)"            "b2(R)"            
#>  [61] "SPS(L)"            "CRE(-ROB,-RUB)(R)" "PB(L1)"           
#>  [64] "NO3(R)"            "AL-VC3l(R)"        "AL-VP1m(R)"       
#>  [67] "PED(R)"            "NO3(L)"            "MB(+ACA)(R)"      
#>  [70] "AL-D(R)"           "AL-VM4(R)"         "NO(R)"            
#>  [73] "BU(R)"             "IPS(R)"            "AL-VP4(R)"        
#>  [76] "CRE(R)"            "PVLP(R)"           "FBl8"             
#>  [79] "LH(R)"             "a'L(L)"            "AL-VA7l(R)"       
#>  [82] "CRE(L)"            "GC"                "AL-VM3(R)"        
#>  [85] "AL-DM4(L)"         "VLNP(R)"           "NO(L)"            
#>  [88] "a'L(R)"            "AL-D(L)"           "AME(R)"           
#>  [91] "SCL(R)"            "AL-DL4(R)"         "AL-VA6(L)"        
#>  [94] "PB(R1)"            "AB(R)"             "AL-VA4(R)"        
#>  [97] "SMP(L)"            "SNP(L)"            "AL-VL2p(R)"       
#> [100] "AOT(R)"            "PB(L5)"            "AMMC"             
#> [103] "PB"                "AL-DM3(L)"         "NO1(R)"           
#> [106] "AL-DA2(L)"         "AL-DA3(L)"         "FBl7"             
#> [109] "GNG"               "AL-DA1(R)"         "AL-DC2(R)"        
#> [112] "AL-VA1d(R)"        "ME(R)"             "bL(R)"            
#> [115] "AL-VA1v(R)"        "SCL(L)"            "CAN(R)"           
#> [118] "EBr5"              "SPS(R)"            "PB(L2)"           
#> [121] "NO"                "AL-VM7d(L)"        "AL-DL2v(R)"       
#> [124] "AL-VL1(R)"         "aL(R)"             "GOR(L)"           
#> [127] "EBr1"              "b1(R)"             "NO2(R)"           
#> [130] "PB(L3)"            "IB112"             "AL-DC4(R)"        
#> [133] "AL-DP1l(R)"        "bL(L)"             "LOP(R)"           
#> [136] "EBr2r4"            "EPA(L)"            "RUB(L)"           
#> [139] "PB(R4)"            "AL-DM5(R)"         "a2(R)"            
#> [142] "AL-VM5v(R)"        "ICL(R)"            "VMNP"             
#> [145] "AL-VP5(R)"         "EBr3am"            "NO1(L)"           
#> [148] "PB(R5)"            "WED(R)"            "EBr3d"            
#> [151] "AL-DP1m(L)"        "AL-DM6(R)"         "AL-VM7v(R)"       
#> [154] "AL-VP1d(R)"        "LAL(-GA)(R)"       "AL-DM2(L)"        
#> [157] "AL-DM3(R)"         "g2(R)"             "FBl6"             
#> [160] "ATL(R)"            "AL(R)"             "AL-DM1(R)"        
#> [163] "AL-VA2(R)"         "AL(L)"             "PB(L9)"           
#> [166] "PB(L8)"            "PB(R9)"            "SAD(-AMMC)"       
#> [169] "PB(R3)"            "AL-VC1(R)"         "g3(R)"            
#> [172] "CX"                "PB(R2)"            "RUB(R)"           
#> [175] "a1(R)"             "AL-DM6(L)"         "PENP"             
#> [178] "BU(L)"             "AL-VP3(R)"         "vACA(R)"          
#> [181] "b'2(R)"            "SIP(R)"            "PB(L6)"           
#> [184] "LX(L)"             "AL-VC3m(R)"        "AL-VM2(R)"        
#> [187] "GF(R)"             "IB"                "GA(R)"            
#> [190] "LX(R)"             "gL(L)"             "AL-DM1(L)"        
#> [193] "INP"               "AL-DL3(R)"         "LAL(L)"           
#> [196] "g5(R)"             "aL(L)"             "AL-DC3(R)"        
#> [199] "AL-DC4(L)"         "AL-VP2(R)"         "EB"               
#> [202] "AVLP(R)"           "AL-DA4l(R)"        "AL-VC5(R)"        
#> [205] "PB(R7)"            "OL(R)"             "AL-DL1(R)"        
#> [208] "AL-VM1(R)"         "AB(L)"             "ATL(L)"           
#> [211] "a'1(R)"            "FLA(R)"            "EPA(R)"           
#> [214] "AL-VM7v(L)"        "EBr3pw"            "PRW"              
#> [217] "FBl1"              "b'L(R)"            "AL-DL4(L)"        
#> [220] "PLP(R)"            "CA(R)"             "AL-VA7m(R)"       
#> [223] "VES(R)"            "SAD"               "a'2(R)"           
#> [226] "AL-DA4m(L)"        "b'L(L)"            "hemibrain"        
#> [229] "AL-VA6(R)"         "FBl9"              "AL-DM5(L)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(L9)"            "VES(L)"            "a2(R)"            
#>   [4] "NO3(L)"            "INP"               "VMNP"             
#>   [7] "EPA(R)"            "RUB(L)"            "CX"               
#>  [10] "SMP(L)"            "GA(R)"             "LAL(-GA)(R)"      
#>  [13] "SPS(L)"            "bL(L)"             "LH(R)"            
#>  [16] "NO1(L)"            "EBr5"              "PB(L4)"           
#>  [19] "PB(L6)"            "EBr3am"            "NO2(L)"           
#>  [22] "NO2(R)"            "b'1(R)"            "g3(R)"            
#>  [25] "AOTU(R)"           "SPS(R)"            "g4(R)"            
#>  [28] "EBr3pw"            "FBl7"              "PB(L2)"           
#>  [31] "FBl4"              "OL(R)"             "NO3(R)"           
#>  [34] "b2(R)"             "EBr1"              "BU(R)"            
#>  [37] "ICL(L)"            "FBl1"              "AME(R)"           
#>  [40] "hemibrain"         "PB(L8)"            "PB(R3)"           
#>  [43] "GF(R)"             "NO1(R)"            "EBr2r4"           
#>  [46] "LO(R)"             "MB(L)"             "SNP(L)"           
#>  [49] "PB(R5)"            "g1(R)"             "dACA(R)"          
#>  [52] "FBl5"              "ICL(R)"            "CAN(R)"           
#>  [55] "PRW"               "CRE(R)"            "SIP(R)"           
#>  [58] "a3(R)"             "AOT(R)"            "PVLP(R)"          
#>  [61] "AVLP(R)"           "MB(R)"             "a'3(R)"           
#>  [64] "FBl8"              "ATL(R)"            "PB(L3)"           
#>  [67] "BU(L)"             "FLA(R)"            "b'L(L)"           
#>  [70] "SMP(R)"            "FB"                "AB(R)"            
#>  [73] "mALT(R)"           "EBr6"              "EBr3d"            
#>  [76] "gL(L)"             "vACA(R)"           "LOP(R)"           
#>  [79] "PED(R)"            "g2(R)"             "PB(R9)"           
#>  [82] "PB(R1)"            "FBl2"              "AB(L)"            
#>  [85] "ME(R)"             "FBl9"              "a'L(L)"           
#>  [88] "aL(R)"             "b'L(R)"            "SAD(-AMMC)"       
#>  [91] "SCL(L)"            "FBl3"              "PB(R4)"           
#>  [94] "bL(R)"             "GOR(R)"            "LAL(L)"           
#>  [97] "lACA(R)"           "AL(L)"             "IPS(R)"           
#> [100] "EPA(L)"            "GNG"               "a1(R)"            
#> [103] "LX(R)"             "gL(R)"             "VES(R)"           
#> [106] "LAL(R)"            "PB(L5)"            "CRE(-ROB,-RUB)(R)"
#> [109] "SLP(R)"            "ATL(L)"            "PB(R6)"           
#> [112] "GC"                "a'1(R)"            "PB(L7)"           
#> [115] "b1(R)"             "aL(L)"             "AMMC"             
#> [118] "CA(L)"             "SAD"               "a'L(R)"           
#> [121] "PB(R2)"            "NO(L)"             "VLNP(R)"          
#> [124] "POC"               "GOR(L)"            "PB(R8)"           
#> [127] "ROB(R)"            "g5(R)"             "CA(R)"            
#> [130] "CRE(L)"            "PENP"              "PB(R7)"           
#> [133] "PB(L1)"            "NO"                "IB"               
#> [136] "NO(R)"             "LX(L)"             "SIP(L)"           
#> [139] "WED(R)"            "SCL(R)"            "PLP(R)"           
#> [142] "a'2(R)"            "EB"                "RUB(R)"           
#> [145] "mALT(L)"           "b'2(R)"            "PB"               
#> [148] "SNP(R)"            "MB(+ACA)(R)"       "FBl6"             
#> [151] "AL(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "NO"                "FBl4"              "ICL(R)"           
#>   [4] "AL-DA4l(R)"        "AL-VP1m(R)"        "AL-D(L)"          
#>   [7] "b'L(L)"            "b'1(R)"            "PB(L1)"           
#>  [10] "AL-VA1d(R)"        "AL-VA5(R)"         "RUB(R)"           
#>  [13] "PENP"              "g1(R)"             "AL-VC1(R)"        
#>  [16] "AL-VM5v(R)"        "mALT(L)"           "AL-DA3(R)"        
#>  [19] "IB"                "ATL(L)"            "gL(L)"            
#>  [22] "AL-DM5(L)"         "AL-VM7d(L)"        "PB(R1)"           
#>  [25] "VES(L)"            "AL-VC5(R)"         "CA(L)"            
#>  [28] "g4(R)"             "AL-DA4m(L)"        "AL-V(R)"          
#>  [31] "PB(L2)"            "PED(R)"            "AL-DL4(L)"        
#>  [34] "lACA(R)"           "GF(R)"             "ROB(R)"           
#>  [37] "CRE(-ROB,-RUB)(R)" "CX"                "PB"               
#>  [40] "EBr2r4"            "EB"                "AL-DC4(L)"        
#>  [43] "AL-DL3(R)"         "AL-VP1d(R)"        "SAD"              
#>  [46] "PLP(R)"            "NO(R)"             "EBr3pw"           
#>  [49] "AL-VA1v(R)"        "AL-DP1m(L)"        "NO3(L)"           
#>  [52] "PB(R2)"            "a'L(L)"            "SLP(R)"           
#>  [55] "MB(+ACA)(R)"       "NO(L)"             "OL(R)"            
#>  [58] "AL-DP1m(R)"        "AL-VP5(R)"         "mALT(R)"          
#>  [61] "AL-VM7v(L)"        "FBl7"              "AL-DP1l(R)"       
#>  [64] "AOTU(R)"           "NO2(R)"            "AL-VA7m(R)"       
#>  [67] "INP"               "AL-VA4(R)"         "AL-VC4(R)"        
#>  [70] "AL-VP4(R)"         "FLA(R)"            "AL-DM6(L)"        
#>  [73] "SAD(-AMMC)"        "LAL(R)"            "AL-VL1(R)"        
#>  [76] "SIP(L)"            "LAL(-GA)(R)"       "FBl8"             
#>  [79] "GOR(R)"            "AL-VC3m(R)"        "AL-VL2p(R)"       
#>  [82] "AL-VM4(R)"         "AL-VP2(R)"         "FBl2"             
#>  [85] "bL(R)"             "AL-VP3(R)"         "GC"               
#>  [88] "SPS(L)"            "g5(R)"             "PB(R3)"           
#>  [91] "PB(L4)"            "AL-VL2a(R)"        "PRW"              
#>  [94] "gL(R)"             "AL-VC3l(R)"        "aL(L)"            
#>  [97] "AL-VM7d(R)"        "hemibrain"         "g2(R)"            
#> [100] "a'2(R)"            "AL-DM3(L)"         "LH(R)"            
#> [103] "IPS(R)"            "SCL(L)"            "EBr3am"           
#> [106] "CRE(R)"            "AL-DM4(L)"         "PB(L6)"           
#> [109] "EPA(R)"            "SIP(R)"            "PB(L7)"           
#> [112] "LO(R)"             "CA(R)"             "AL-DA3(L)"        
#> [115] "FBl3"              "AL(R)"             "AL-DA4m(R)"       
#> [118] "AL-DM6(R)"         "EBr5"              "AL-VM7v(R)"       
#> [121] "POC"               "a'1(R)"            "AL-DA1(R)"        
#> [124] "AL-DC4(R)"         "a'3(R)"            "PB(R6)"           
#> [127] "SCL(R)"            "AL-VC2(R)"         "AL-VM3(R)"        
#> [130] "PB(R7)"            "NO1(L)"            "AL(L)"            
#> [133] "ICL(L)"            "FBl6"              "a3(R)"            
#> [136] "AL-DC2(L)"         "CRE(L)"            "FBl1"             
#> [139] "FBl9"              "AL-VA3(R)"         "AL-VA7l(R)"       
#> [142] "EBr1"              "PB(R4)"            "AL-DC1(R)"        
#> [145] "AL-DL5(L)"         "AL-DM1(R)"         "AL-DM3(R)"        
#> [148] "SPS(R)"            "NO3(R)"            "AVLP(R)"          
#> [151] "AME(R)"            "AL-DL4(R)"         "AL-DM4(R)"        
#> [154] "a2(R)"             "PB(L8)"            "SNP(R)"           
#> [157] "AL-VM1(R)"         "WED(R)"            "FB"               
#> [160] "dACA(R)"           "VES(R)"            "GOR(L)"           
#> [163] "a'L(R)"            "CAN(R)"            "VLNP(R)"          
#> [166] "LOP(R)"            "RUB(L)"            "AL-VM5d(R)"       
#> [169] "NO2(L)"            "SMP(L)"            "MB(L)"            
#> [172] "AL-DC3(R)"         "FB-column3"        "AL-DM5(R)"        
#> [175] "aL(R)"             "AL-DA2(L)"         "AL-DC1(L)"        
#> [178] "a1(R)"             "EBr6"              "LX(R)"            
#> [181] "AL-VA2(R)"         "AL-VP1l(R)"        "vACA(R)"          
#> [184] "BU(R)"             "AL-DM2(L)"         "CRE(-RUB)(L)"     
#> [187] "g3(R)"             "PB(R5)"            "FBl5"             
#> [190] "b1(R)"             "EPA(L)"            "BU(L)"            
#> [193] "PB(L9)"            "LX(L)"             "AOT(R)"           
#> [196] "AL-DL2v(R)"        "PB(L3)"            "b2(R)"            
#> [199] "PB(R8)"            "EBr3d"             "ATL(R)"           
#> [202] "AL-DA2(R)"         "AL-VA6(L)"         "bL(L)"            
#> [205] "AL-D(R)"           "AL-VA6(R)"         "PVLP(R)"          
#> [208] "AL-VM2(R)"         "SMP(R)"            "GA(R)"            
#> [211] "NO1(R)"            "VMNP"              "ME(R)"            
#> [214] "AL-DL1(R)"         "AL-DM2(R)"         "b'L(R)"           
#> [217] "GNG"               "AL-DC2(R)"         "AL-DL2d(R)"       
#> [220] "AB(L)"             "PB(R9)"            "AB(R)"            
#> [223] "MB(R)"             "PB(L5)"            "AMMC"             
#> [226] "AL-DM1(L)"         "b'2(R)"            "LAL(L)"           
#> [229] "SNP(L)"            "AL-DL5(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "b1(R)"             "PB(L4)"            "PB(R5)"           
#>   [4] "VMNP"              "OL(R)"             "AL-DM1(R)"        
#>   [7] "AL-VP1d(R)"        "GOR(L)"            "AL-D(L)"          
#>  [10] "AL-DM1(L)"         "PB(L7)"            "SNP(L)"           
#>  [13] "GOR(R)"            "SCL(R)"            "EBr2r4"           
#>  [16] "a'L(R)"            "AL-DC1(L)"         "EBr3pw"           
#>  [19] "LO(R)"             "AL-DL2v(R)"        "AL-VC4(R)"        
#>  [22] "SCL(L)"            "a'3(R)"            "AOTU(R)"          
#>  [25] "a2(R)"             "g1(R)"             "FBl7"             
#>  [28] "NO(L)"             "AL-DA3(R)"         "AL-VL1(R)"        
#>  [31] "a1(R)"             "CRE(-RUB)(L)"      "AL-DP1l(R)"       
#>  [34] "aL(R)"             "LX(R)"             "CA(R)"            
#>  [37] "SLP(R)"            "PB(L5)"            "EBr5"             
#>  [40] "EB"                "LX(L)"             "AL-DA4m(R)"       
#>  [43] "AL-VC1(R)"         "PB(R8)"            "AL-DA2(R)"        
#>  [46] "AL-DL1(R)"         "AL-DM5(R)"         "AL-VA6(R)"        
#>  [49] "AL-VL2p(R)"        "b'1(R)"            "AMMC"             
#>  [52] "AL-VM7v(L)"        "NO3(L)"            "RUB(R)"           
#>  [55] "AL-DC4(L)"         "AL-DP1m(L)"        "PB(R1)"           
#>  [58] "AL-DL3(R)"         "AL-DA3(L)"         "gL(L)"            
#>  [61] "ROB(R)"            "AL-DP1m(R)"        "AL-VC3m(R)"       
#>  [64] "AL-V(R)"           "VLNP(R)"           "dACA(R)"          
#>  [67] "ME(R)"             "g4(R)"             "PB(L9)"           
#>  [70] "CRE(R)"            "AL-DA2(L)"         "FBl1"             
#>  [73] "LH(R)"             "PB(R3)"            "CX"               
#>  [76] "IPS(R)"            "AL-DC2(R)"         "SMP(R)"           
#>  [79] "GF(R)"             "MB(L)"             "BU(L)"            
#>  [82] "PVLP(R)"           "PB(L3)"            "LAL(-GA)(R)"      
#>  [85] "AL-DC3(R)"         "AL-VA3(R)"         "AL-VL2a(R)"       
#>  [88] "FBl8"              "AL-VM1(R)"         "g3(R)"            
#>  [91] "lACA(R)"           "PB(L8)"            "AB(L)"            
#>  [94] "FB-column3"        "NO1(R)"            "BU(R)"            
#>  [97] "FBl9"              "b'L(L)"            "mALT(R)"          
#> [100] "PB(L6)"            "gL(R)"             "EPA(L)"           
#> [103] "EBr3d"             "AOT(R)"            "NO2(L)"           
#> [106] "AL-DM6(L)"         "AL-VA6(L)"         "FBl2"             
#> [109] "FBl4"              "AL-VC5(R)"         "vACA(R)"          
#> [112] "bL(L)"             "AL-VP1l(R)"        "MB(R)"            
#> [115] "WED(R)"            "PB(L1)"            "AL-DC1(R)"        
#> [118] "AL-DL5(R)"         "AL-VM5d(R)"        "g2(R)"            
#> [121] "AL-DM3(L)"         "AL-D(R)"           "AL-VP5(R)"        
#> [124] "EBr3am"            "SAD"               "SIP(R)"           
#> [127] "NO"                "NO(R)"             "a3(R)"            
#> [130] "VES(L)"            "AL-VM3(R)"         "PLP(R)"           
#> [133] "AL-DC2(L)"         "a'1(R)"            "LAL(R)"           
#> [136] "AL-VA7l(R)"        "AL-VC3l(R)"        "GA(R)"            
#> [139] "NO3(R)"            "NO1(L)"            "b2(R)"            
#> [142] "FB"                "ICL(L)"            "AL-DA4l(R)"       
#> [145] "AL-DM2(R)"         "AL-VA7m(R)"        "AL-VP4(R)"        
#> [148] "PB(R6)"            "FLA(R)"            "VES(R)"           
#> [151] "FBl3"              "AL-DM4(R)"         "AL-VP1m(R)"       
#> [154] "PB"                "MB(+ACA)(R)"       "AL-DL4(L)"        
#> [157] "INP"               "AL-VA4(R)"         "SPS(R)"           
#> [160] "CAN(R)"            "SMP(L)"            "AL(L)"            
#> [163] "a'L(L)"            "AL-DA1(R)"         "b'2(R)"           
#> [166] "AME(R)"            "PRW"               "GC"               
#> [169] "a'2(R)"            "g5(R)"             "AL-DC4(R)"        
#> [172] "CRE(L)"            "FBl5"              "AL-VM2(R)"        
#> [175] "AL-VM5v(R)"        "SIP(L)"            "aL(L)"            
#> [178] "AL-VA2(R)"         "PB(R2)"            "RUB(L)"           
#> [181] "AL-VA5(R)"         "AL-VC2(R)"         "AL-VP2(R)"        
#> [184] "POC"               "EPA(R)"            "FBl6"             
#> [187] "AL-VP3(R)"         "LOP(R)"            "AL-DM2(L)"        
#> [190] "IB"                "AB(R)"             "GNG"              
#> [193] "bL(R)"             "AL-VA1v(R)"        "PED(R)"           
#> [196] "AL-DL5(L)"         "CRE(-ROB,-RUB)(R)" "SPS(L)"           
#> [199] "NO2(R)"            "AL-DL4(R)"         "b'L(R)"           
#> [202] "EBr6"              "AL(R)"             "AL-VM7d(R)"       
#> [205] "SNP(R)"            "ATL(R)"            "PB(L2)"           
#> [208] "PB(R7)"            "AL-DM5(L)"         "AL-VM7v(R)"       
#> [211] "ATL(L)"            "AVLP(R)"           "AL-DM3(R)"        
#> [214] "AL-DM4(L)"         "PB(R9)"            "SAD(-AMMC)"       
#> [217] "AL-DM6(R)"         "AL-VA1d(R)"        "LAL(L)"           
#> [220] "hemibrain"         "AL-VM4(R)"         "PENP"             
#> [223] "ICL(R)"            "AL-DL2d(R)"        "CA(L)"            
#> [226] "AL-DA4m(L)"        "AL-VM7d(L)"        "mALT(L)"          
#> [229] "EBr1"              "PB(R4)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method   path                                      
#> [1,] "PATCH"  "/api/*"                                  
#> [2,] "GET"    "/api/v:ver/cached/roiconnectivity"       
#> [3,] "DELETE" "/api"                                    
#> [4,] "GET"    "/api/v:ver/custom/custom"                
#> [5,] "POST"   "/api/v:ver/npexplorer/commonconnectivity"
#> [6,] "GET"    "/api/skeletons/skeleton/:dataset/:id"    
#>      name                                                                                
#> [1,] "github.com/labstack/echo/v4.glob..func1"                                           
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                
#> [3,] "github.com/labstack/echo/v4.glob..func1"                                           
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/skeletons.masterAPI.getSkeleton-fm"
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }