Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DM6(L)"         "b'1(R)"            "AL-DC4(R)"        
#>   [4] "AL-DL2d(R)"        "SCL(L)"            "AL-VC5(R)"        
#>   [7] "AL-VM7d(R)"        "g1(R)"             "ME(R)"            
#>  [10] "ROB(R)"            "PB(L9)"            "b1(R)"            
#>  [13] "CRE(R)"            "AL-DL4(L)"         "ICL(L)"           
#>  [16] "EBr5"              "EBr3d"             "bL(L)"            
#>  [19] "aL(L)"             "AL-VM5d(R)"        "AL-DA2(L)"        
#>  [22] "AL-DP1m(L)"        "AL-VA1v(R)"        "IB"               
#>  [25] "CA(L)"             "b'2(R)"            "AL-D(L)"          
#>  [28] "VLNP(R)"           "a'L(R)"            "PB(L7)"           
#>  [31] "a2(R)"             "NO1(R)"            "EPA(L)"           
#>  [34] "AL-DA4l(R)"        "CRE(L)"            "PB(R3)"           
#>  [37] "AL-DA2(R)"         "AL-VA2(R)"         "MB(L)"            
#>  [40] "EBr6"              "AL-DL5(L)"         "LX(L)"            
#>  [43] "ATL(L)"            "PB(L8)"            "FBl1"             
#>  [46] "PB(R4)"            "WED(R)"            "VES(R)"           
#>  [49] "AL-VA5(R)"         "AL-VM3(R)"         "LAL(L)"           
#>  [52] "AL-VA1d(R)"        "SNP(L)"            "AMMC"             
#>  [55] "OL(R)"             "AL-DM1(L)"         "AL-DC1(R)"        
#>  [58] "EBr3am"            "AL-VM5v(R)"        "mALT(R)"          
#>  [61] "gL(L)"             "AL-DM5(L)"         "AVLP(R)"          
#>  [64] "PLP(R)"            "AL-VM7v(L)"        "AL-DL1(R)"        
#>  [67] "POC"               "RUB(R)"            "PB(L6)"           
#>  [70] "AL-VC1(R)"         "AL-DA3(L)"         "AL-VC2(R)"        
#>  [73] "GC"                "PB(R1)"            "a'L(L)"           
#>  [76] "AL-DM6(R)"         "VES(L)"            "PB(L3)"           
#>  [79] "EBr1"              "AL-VA6(L)"         "FBl5"             
#>  [82] "FBl8"              "AL-VP4(R)"         "PB(R7)"           
#>  [85] "PB"                "AL-VM7d(L)"        "SIP(R)"           
#>  [88] "AL(R)"             "AOTU(R)"           "LO(R)"            
#>  [91] "SLP(R)"            "AL-DA1(R)"         "AL-DA4m(R)"       
#>  [94] "GF(R)"             "EPA(R)"            "IB112"            
#>  [97] "AL-DM3(R)"         "AL-VL2p(R)"        "SCL(R)"           
#> [100] "AL-DC2(L)"         "AL-DM4(R)"         "AL-DM5(R)"        
#> [103] "NO1(L)"            "b2(R)"             "FBl4"             
#> [106] "AL-VC4(R)"         "AL-VA3(R)"         "AL-DC1(L)"        
#> [109] "a'1(R)"            "FBl9"              "AL-VA7l(R)"       
#> [112] "AL-VP2(R)"         "PED(R)"            "FLA(R)"           
#> [115] "AL-VC3m(R)"        "a3(R)"             "AL-VA6(R)"        
#> [118] "mALT(L)"           "hemibrain"         "SNP(R)"           
#> [121] "EB"                "AME(R)"            "GNG"              
#> [124] "FBl7"              "NO2(L)"            "g3(R)"            
#> [127] "BU(L)"             "PB(R8)"            "AL(L)"            
#> [130] "AL-DM3(L)"         "PB(R6)"            "AL-VM4(R)"        
#> [133] "AB(L)"             "AL-DL2v(R)"        "PVLP(R)"          
#> [136] "FBl3"              "AL-DC2(R)"         "AL-DP1m(R)"       
#> [139] "AL-DM2(L)"         "LOP(R)"            "LAL(-GA)(R)"      
#> [142] "CRE(-RUB)(L)"      "FB-column3"        "AL-DA3(R)"        
#> [145] "GA(R)"             "NO3(R)"            "AL-DL3(R)"        
#> [148] "AL-VP3(R)"         "MB(+ACA)(R)"       "FBl2"             
#> [151] "PB(R2)"            "RUB(L)"            "AL-VP5(R)"        
#> [154] "PB(L2)"            "FB"                "SAD(-AMMC)"       
#> [157] "NO(L)"             "AL-DL4(R)"         "b'L(R)"           
#> [160] "PB(L1)"            "AL-VC3l(R)"        "a'3(R)"           
#> [163] "AL-VA4(R)"         "AOT(R)"            "vACA(R)"          
#> [166] "VMNP"              "PB(R9)"            "NO3(L)"           
#> [169] "AL-DC3(R)"         "AL-D(R)"           "PB(L5)"           
#> [172] "SMP(L)"            "SPS(L)"            "NO"               
#> [175] "NO(R)"             "SMP(R)"            "GOR(L)"           
#> [178] "AL-DM4(L)"         "INP"               "AL-DM1(R)"        
#> [181] "AL-VM2(R)"         "dACA(R)"           "EBr3pw"           
#> [184] "FBl6"              "IPS(R)"            "AL-VP1m(R)"       
#> [187] "SPS(R)"            "LX(R)"             "g5(R)"            
#> [190] "AL-VL2a(R)"        "AL-DC4(L)"         "CA(R)"            
#> [193] "AL-VA7m(R)"        "PRW"               "LH(R)"            
#> [196] "a1(R)"             "EBr2r4"            "GOR(R)"           
#> [199] "AL-VP1l(R)"        "PB(R5)"            "CRE(-ROB,-RUB)(R)"
#> [202] "gL(R)"             "LAL(R)"            "AL-DP1l(R)"       
#> [205] "AL-VM7v(R)"        "BU(R)"             "a'2(R)"           
#> [208] "AL-DA4m(L)"        "bL(R)"             "AL-DM2(R)"        
#> [211] "PENP"              "lACA(R)"           "g4(R)"            
#> [214] "ICL(R)"            "b'L(L)"            "PB(L4)"           
#> [217] "MB(R)"             "AL-VL1(R)"         "CX"               
#> [220] "SAD"               "SIP(L)"            "aL(R)"            
#> [223] "CAN(R)"            "AL-V(R)"           "g2(R)"            
#> [226] "ATL(R)"            "AL-DL5(R)"         "AL-VM1(R)"        
#> [229] "AB(R)"             "NO2(R)"            "AL-VP1d(R)"       
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "NO"                "bL(R)"             "g3(R)"            
#>   [4] "mALT(R)"           "PB(R9)"            "SNP(L)"           
#>   [7] "PVLP(R)"           "GNG"               "GC"               
#>  [10] "CRE(R)"            "g2(R)"             "vACA(R)"          
#>  [13] "b'2(R)"            "b'L(R)"            "NO(R)"            
#>  [16] "EBr3pw"            "PB(R5)"            "mALT(L)"          
#>  [19] "PB(R8)"            "EBr2r4"            "FBl9"             
#>  [22] "PB(L2)"            "NO1(L)"            "a'1(R)"           
#>  [25] "LOP(R)"            "g5(R)"             "SAD(-AMMC)"       
#>  [28] "VLNP(R)"           "SMP(R)"            "CX"               
#>  [31] "a2(R)"             "SCL(R)"            "EBr5"             
#>  [34] "BU(R)"             "bL(L)"             "FBl8"             
#>  [37] "PENP"              "SIP(L)"            "lACA(R)"          
#>  [40] "EPA(R)"            "EBr6"              "LX(R)"            
#>  [43] "PB(L9)"            "LO(R)"             "GOR(R)"           
#>  [46] "AMMC"              "IB"                "SMP(L)"           
#>  [49] "hemibrain"         "EBr1"              "CRE(L)"           
#>  [52] "PB(L7)"            "LAL(R)"            "NO2(R)"           
#>  [55] "CRE(-ROB,-RUB)(R)" "FLA(R)"            "VMNP"             
#>  [58] "SNP(R)"            "ICL(L)"            "CA(R)"            
#>  [61] "ICL(R)"            "PB(L4)"            "OL(R)"            
#>  [64] "PED(R)"            "LAL(-GA)(R)"       "PB(L1)"           
#>  [67] "ME(R)"             "PB(L6)"            "FBl1"             
#>  [70] "MB(L)"             "LX(L)"             "AB(L)"            
#>  [73] "FBl6"              "aL(L)"             "EBr3am"           
#>  [76] "PB(R2)"            "EPA(L)"            "PB(R6)"           
#>  [79] "SPS(R)"            "g1(R)"             "FB"               
#>  [82] "dACA(R)"           "FBl2"              "FBl4"             
#>  [85] "AVLP(R)"           "b'L(L)"            "NO(L)"            
#>  [88] "AOT(R)"            "SLP(R)"            "AL(R)"            
#>  [91] "SAD"               "b2(R)"             "FBl3"             
#>  [94] "b'1(R)"            "AOTU(R)"           "PB(R1)"           
#>  [97] "SCL(L)"            "GOR(L)"            "IPS(R)"           
#> [100] "a3(R)"             "PB(L3)"            "aL(R)"            
#> [103] "EB"                "NO3(L)"            "PB(R3)"           
#> [106] "ROB(R)"            "CAN(R)"            "POC"              
#> [109] "RUB(R)"            "PB(L5)"            "ATL(L)"           
#> [112] "PB(R7)"            "EBr3d"             "g4(R)"            
#> [115] "SIP(R)"            "FBl7"              "b1(R)"            
#> [118] "a1(R)"             "a'L(L)"            "NO1(R)"           
#> [121] "SPS(L)"            "PLP(R)"            "gL(R)"            
#> [124] "MB(+ACA)(R)"       "ATL(R)"            "PB"               
#> [127] "VES(L)"            "a'3(R)"            "LAL(L)"           
#> [130] "GA(R)"             "PB(L8)"            "AL(L)"            
#> [133] "MB(R)"             "CA(L)"             "GF(R)"            
#> [136] "WED(R)"            "NO3(R)"            "BU(L)"            
#> [139] "a'L(R)"            "FBl5"              "NO2(L)"           
#> [142] "PRW"               "gL(L)"             "LH(R)"            
#> [145] "AB(R)"             "INP"               "PB(R4)"           
#> [148] "a'2(R)"            "VES(R)"            "AME(R)"           
#> [151] "RUB(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "PB(R6)"            "AL-VC3m(R)"        "SIP(L)"           
#>   [4] "AL-VP4(R)"         "AL-VA3(R)"         "AL-VA6(R)"        
#>   [7] "AL-VC1(R)"         "PB(L6)"            "MB(R)"            
#>  [10] "g3(R)"             "lACA(R)"           "AL-VL2a(R)"       
#>  [13] "SMP(R)"            "b'1(R)"            "AL-DL5(L)"        
#>  [16] "AL-DA4l(R)"        "AL-DC1(R)"         "AL-DC3(R)"        
#>  [19] "PB(L3)"            "PB(L1)"            "a'1(R)"           
#>  [22] "ATL(R)"            "CA(L)"             "AL-DC1(L)"        
#>  [25] "PB(R9)"            "ROB(R)"            "AL-VA1v(R)"       
#>  [28] "CRE(-ROB,-RUB)(R)" "EBr1"              "GOR(R)"           
#>  [31] "AL-DA2(R)"         "AL-VM2(R)"         "BU(R)"            
#>  [34] "AL-DM6(L)"         "SNP(L)"            "b'L(L)"           
#>  [37] "ATL(L)"            "MB(L)"             "a'L(R)"           
#>  [40] "PB(L7)"            "AL-DC4(R)"         "g4(R)"            
#>  [43] "AL-DM1(L)"         "AL-VC4(R)"         "g1(R)"            
#>  [46] "VMNP"              "a'2(R)"            "EB"               
#>  [49] "IB"                "mALT(L)"           "GA(R)"            
#>  [52] "EBr6"              "AL(R)"             "AL-VL1(R)"        
#>  [55] "PRW"               "AL-DL5(R)"         "hemibrain"        
#>  [58] "MB(+ACA)(R)"       "FBl2"              "AL-DA3(R)"        
#>  [61] "AL-VP2(R)"         "LAL(-GA)(R)"       "CRE(-RUB)(L)"     
#>  [64] "VES(L)"            "AL-VM4(R)"         "PB(L4)"           
#>  [67] "AL-V(R)"           "EBr5"              "AL-DM3(R)"        
#>  [70] "AL-DP1m(R)"        "AL-VC2(R)"         "AL-VA1d(R)"       
#>  [73] "AL-VP3(R)"         "FBl6"              "PB(R4)"           
#>  [76] "OL(R)"             "AL-DM2(R)"         "EBr2r4"           
#>  [79] "AME(R)"            "AL-DM1(R)"         "AL-VP1l(R)"       
#>  [82] "a'3(R)"            "SAD"               "FB"               
#>  [85] "AL-VL2p(R)"        "AB(L)"             "AL-DM3(L)"        
#>  [88] "INP"               "AL-DL2d(R)"        "CX"               
#>  [91] "AL-DC4(L)"         "NO(R)"             "FBl4"             
#>  [94] "AL-VM5v(R)"        "PENP"              "LAL(L)"           
#>  [97] "g2(R)"             "AL-DA4m(L)"        "SCL(L)"           
#> [100] "mALT(R)"           "AL-DL4(L)"         "bL(R)"            
#> [103] "RUB(L)"            "AL-VA7l(R)"        "NO3(R)"           
#> [106] "AL(L)"             "FBl9"              "ICL(R)"           
#> [109] "PB(L2)"            "LX(R)"             "AL-DC2(L)"        
#> [112] "NO2(R)"            "AL-DA4m(R)"        "GC"               
#> [115] "SPS(L)"            "GNG"               "NO1(R)"           
#> [118] "aL(L)"             "AL-VC5(R)"         "a2(R)"            
#> [121] "WED(R)"            "AL-VM7d(L)"        "PB(R1)"           
#> [124] "EBr3pw"            "AL-D(R)"           "b2(R)"            
#> [127] "SCL(R)"            "AL-DA2(L)"         "AL-VP5(R)"        
#> [130] "PED(R)"            "NO3(L)"            "PB(R3)"           
#> [133] "AL-VA5(R)"         "AL-DP1m(L)"        "AL-VA6(L)"        
#> [136] "g5(R)"             "RUB(R)"            "NO"               
#> [139] "LAL(R)"            "AL-VM1(R)"         "AL-VA2(R)"        
#> [142] "AL-VM3(R)"         "AL-VM5d(R)"        "AL-D(L)"          
#> [145] "LX(L)"             "AB(R)"             "PLP(R)"           
#> [148] "SNP(R)"            "bL(L)"             "FBl7"             
#> [151] "AOTU(R)"           "SPS(R)"            "a'L(L)"           
#> [154] "FBl8"              "PB(R2)"            "AL-VA4(R)"        
#> [157] "CRE(L)"            "SAD(-AMMC)"        "AL-DL3(R)"        
#> [160] "AL-VP1m(R)"        "NO2(L)"            "b'2(R)"           
#> [163] "NO1(L)"            "gL(R)"             "AL-DA1(R)"        
#> [166] "aL(R)"             "AL-DM5(L)"         "EPA(L)"           
#> [169] "a3(R)"             "VES(R)"            "PVLP(R)"          
#> [172] "AL-DC2(R)"         "PB(R5)"            "GOR(L)"           
#> [175] "NO(L)"             "AL-VC3l(R)"        "SMP(L)"           
#> [178] "b'L(R)"            "PB(R8)"            "LO(R)"            
#> [181] "PB"                "ICL(L)"            "SLP(R)"           
#> [184] "POC"               "GF(R)"             "EBr3d"            
#> [187] "AL-DA3(L)"         "AL-DM2(L)"         "FLA(R)"           
#> [190] "IPS(R)"            "AL-VP1d(R)"        "PB(L5)"           
#> [193] "PB(R7)"            "CRE(R)"            "PB(L8)"           
#> [196] "AL-VM7v(R)"        "LH(R)"             "b1(R)"            
#> [199] "BU(L)"             "AL-DP1l(R)"        "ME(R)"            
#> [202] "AL-DL2v(R)"        "AL-VM7d(R)"        "EBr3am"           
#> [205] "AMMC"              "SIP(R)"            "FBl1"             
#> [208] "FBl3"              "FBl5"              "vACA(R)"          
#> [211] "AL-DM4(L)"         "AL-DL4(R)"         "AL-DM4(R)"        
#> [214] "VLNP(R)"           "PB(L9)"            "CAN(R)"           
#> [217] "AL-DM6(R)"         "AL-DL1(R)"         "a1(R)"            
#> [220] "dACA(R)"           "AL-VM7v(L)"        "AVLP(R)"          
#> [223] "AL-DM5(R)"         "gL(L)"             "EPA(R)"           
#> [226] "LOP(R)"            "FB-column3"        "CA(R)"            
#> [229] "AL-VA7m(R)"        "AOT(R)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "CRE(R)"            "ATL(L)"            "CA(L)"            
#>   [4] "AL-VM4(R)"         "AL-DL4(L)"         "ICL(L)"           
#>   [7] "CAN(R)"            "NO1(R)"            "g4(R)"            
#>  [10] "AL-VP1m(R)"        "EBr5"              "NO2(R)"           
#>  [13] "PB(R4)"            "NO1(L)"            "a3(R)"            
#>  [16] "AL-DM6(R)"         "AL-VP1l(R)"        "AOTU(R)"          
#>  [19] "aL(R)"             "FB-column3"        "LAL(L)"           
#>  [22] "GF(R)"             "CRE(L)"            "SPS(L)"           
#>  [25] "AL-DP1m(R)"        "AL-VC3m(R)"        "FBl3"             
#>  [28] "MB(R)"             "AL-VA1d(R)"        "VMNP"             
#>  [31] "NO(R)"             "NO(L)"             "AL-VA5(R)"        
#>  [34] "AL-VM2(R)"         "a'L(L)"            "SLP(R)"           
#>  [37] "SNP(R)"            "AL-VA6(L)"         "AL-VM7v(L)"       
#>  [40] "AL-DA2(L)"         "NO"                "AL-DA4m(R)"       
#>  [43] "LH(R)"             "bL(R)"             "RUB(L)"           
#>  [46] "VES(L)"            "AL-DM2(R)"         "AL-VA4(R)"        
#>  [49] "LX(R)"             "CRE(-RUB)(L)"      "b1(R)"            
#>  [52] "vACA(R)"           "a'L(R)"            "AL-D(L)"          
#>  [55] "PB(R3)"            "AB(R)"             "AL(R)"            
#>  [58] "AL-DC2(L)"         "AL-DM2(L)"         "NO3(L)"           
#>  [61] "FBl1"              "AL-VC2(R)"         "VLNP(R)"          
#>  [64] "mALT(R)"           "EBr1"              "a'1(R)"           
#>  [67] "PB(R9)"            "ATL(R)"            "AMMC"             
#>  [70] "WED(R)"            "INP"               "AL-DL3(R)"        
#>  [73] "AL-DM4(R)"         "FBl5"              "FBl8"             
#>  [76] "EBr2r4"            "AL-DC4(L)"         "FBl2"             
#>  [79] "b'L(R)"            "FBl7"              "PB(L2)"           
#>  [82] "SIP(R)"            "AL-DC1(R)"         "PB(R5)"           
#>  [85] "lACA(R)"           "PB(L8)"            "PLP(R)"           
#>  [88] "AL-DA4m(L)"        "AL-DA3(R)"         "AOT(R)"           
#>  [91] "SMP(L)"            "g2(R)"             "PB(L5)"           
#>  [94] "FB"                "PB(R8)"            "AL-VA3(R)"        
#>  [97] "PB"                "SAD(-AMMC)"        "a'3(R)"           
#> [100] "b'1(R)"            "PB(L3)"            "PB(L1)"           
#> [103] "AL-VP4(R)"         "AL-DM5(R)"         "AL-VL1(R)"        
#> [106] "LAL(-GA)(R)"       "b2(R)"             "FBl4"             
#> [109] "PB(R1)"            "AL-VC4(R)"         "AL-VM1(R)"        
#> [112] "AL-VP3(R)"         "SPS(R)"            "BU(R)"            
#> [115] "AL-DM6(L)"         "SCL(R)"            "POC"              
#> [118] "AL-DM3(L)"         "AL-DL5(R)"         "AL-VP2(R)"        
#> [121] "IPS(R)"            "AL-DC3(R)"         "AB(L)"            
#> [124] "gL(R)"             "MB(+ACA)(R)"       "PB(R2)"           
#> [127] "AL-D(R)"           "AL-DM5(L)"         "PVLP(R)"          
#> [130] "PB(L4)"            "SCL(L)"            "EBr3am"           
#> [133] "PB(R7)"            "MB(L)"             "EBr3d"            
#> [136] "AL-VP5(R)"         "EB"                "FBl6"             
#> [139] "AL-VA6(R)"         "mALT(L)"           "LAL(R)"           
#> [142] "OL(R)"             "AL-VA1v(R)"        "FLA(R)"           
#> [145] "b'2(R)"            "SAD"               "BU(L)"            
#> [148] "AL-VA7m(R)"        "PB(L7)"            "AL-VL2p(R)"       
#> [151] "AL-VM3(R)"         "AL-DM1(R)"         "CRE(-ROB,-RUB)(R)"
#> [154] "AL-DL5(L)"         "FBl9"              "AL-DA1(R)"        
#> [157] "hemibrain"         "GOR(L)"            "AL-DP1m(L)"       
#> [160] "EBr3pw"            "GOR(R)"            "CX"               
#> [163] "SIP(L)"            "g1(R)"             "EBr6"             
#> [166] "GNG"               "AL-DL2d(R)"        "AL-DL2v(R)"       
#> [169] "PB(R6)"            "a'2(R)"            "ICL(R)"           
#> [172] "PB(L9)"            "SNP(L)"            "a1(R)"            
#> [175] "NO2(L)"            "AL-DM3(R)"         "AL-VP1d(R)"       
#> [178] "SMP(R)"            "AL-DA4l(R)"        "g3(R)"            
#> [181] "GA(R)"             "AL-DC1(L)"         "gL(L)"            
#> [184] "AL-DM1(L)"         "AL-VC3l(R)"        "ROB(R)"           
#> [187] "bL(L)"             "AL-VM7d(R)"        "AL-VC5(R)"        
#> [190] "RUB(R)"            "AL-VA2(R)"         "AL-VM5d(R)"       
#> [193] "PRW"               "PB(L6)"            "EPA(L)"           
#> [196] "AL-DA3(L)"         "b'L(L)"            "AL-VC1(R)"        
#> [199] "LOP(R)"            "dACA(R)"           "AL(L)"            
#> [202] "AME(R)"            "AL-VM7v(R)"        "AL-V(R)"          
#> [205] "AL-DM4(L)"         "aL(L)"             "AL-DL1(R)"        
#> [208] "IB"                "PENP"              "VES(R)"           
#> [211] "AL-DP1l(R)"        "ME(R)"             "EPA(R)"           
#> [214] "AL-DC2(R)"         "AL-VL2a(R)"        "AL-VM5v(R)"       
#> [217] "a2(R)"             "PED(R)"            "GC"               
#> [220] "g5(R)"             "AVLP(R)"           "LX(L)"            
#> [223] "LO(R)"             "AL-DL4(R)"         "AL-VA7l(R)"       
#> [226] "CA(R)"             "AL-DA2(R)"         "AL-DC4(R)"        
#> [229] "NO3(R)"            "AL-VM7d(L)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`manc:v1.0`
#> $`manc:v1.0`$`last-mod`
#> [1] "2023-05-31 23:54:01"
#> 
#> $`manc:v1.0`$uuid
#> [1] "59b37970bc7a4341b9a3a965a0d6b402"
#> 
#> $`manc:v1.0`$ROIs
#>  [1] "AbN1(L)"          "AbN2(R)"          "VProN(L)"         "AbN3(R)"         
#>  [5] "mVAC(T3)(L)"      "Ov(L)"            "ADMN(R)"          "MesoAN(R)"       
#>  [9] "mVAC(T1)(L)"      "NTct(UTct-T1)(R)" "WTct(UTct-T2)(R)" "GF(L)"           
#> [13] "LegNp(T1)(L)"     "mVAC(T2)(R)"      "MesoLN(R)"        "ProCN(L)"        
#> [17] "HTct(UTct-T3)(R)" "PrN(L)"           "LegNp(T2)(R)"     "mVAC(T3)(R)"     
#> [21] "CvN(L)"           "DProN(L)"         "MesoLN(L)"        "ProAN(L)"        
#> [25] "GF(R)"            "LegNp(T3)(R)"     "ProCN(R)"         "AbN3(L)"         
#> [29] "AbNT"             "MetaLN(L)"        "VProN(R)"         "LegNp(T3)(L)"    
#> [33] "AbN1(R)"          "AbN4(R)"          "ProLN(R)"         "DMetaN(R)"       
#> [37] "CV"               "mVAC(T1)(R)"      "mVAC(T2)(L)"      "NTct(UTct-T1)(L)"
#> [41] "HTct(UTct-T3)(L)" "ADMN(L)"          "DMetaN(L)"        "MesoAN(L)"       
#> [45] "PrN(R)"           "Ov(R)"            "ANm"              "WTct(UTct-T2)(L)"
#> [49] "AbN2(L)"          "ProLN(L)"         "LegNp(T1)(R)"     "CvN(R)"          
#> [53] "MetaLN(R)"        "PDMN(L)"          "ProAN(R)"         "LegNp(T2)(L)"    
#> [57] "IntTct"           "LTct"             "AbN4(L)"          "DProN(R)"        
#> [61] "PDMN(R)"         
#> 
#> $`manc:v1.0`$superLevelROIs
#>  [1] "ADMN(L)"          "ADMN(R)"          "Ov(L)"            "Ov(R)"           
#>  [5] "ANm"              "AbN1(L)"          "AbN1(R)"          "AbN2(L)"         
#>  [9] "AbN2(R)"          "AbN3(L)"          "AbN3(R)"          "AbN4(L)"         
#> [13] "AbN4(R)"          "AbNT"             "CV"               "CvN(L)"          
#> [17] "CvN(R)"           "DMetaN(L)"        "DMetaN(R)"        "DProN(L)"        
#> [21] "DProN(R)"         "GF(L)"            "GF(R)"            "HTct(UTct-T3)(L)"
#> [25] "HTct(UTct-T3)(R)" "LegNp(T1)(L)"     "LegNp(T1)(R)"     "LegNp(T2)(L)"    
#> [29] "LegNp(T2)(R)"     "LegNp(T3)(L)"     "LegNp(T3)(R)"     "IntTct"          
#> [33] "LTct"             "MesoAN(L)"        "MesoAN(R)"        "MesoLN(L)"       
#> [37] "MesoLN(R)"        "MetaLN(L)"        "MetaLN(R)"        "NTct(UTct-T1)(L)"
#> [41] "NTct(UTct-T1)(R)" "PDMN(L)"          "PrN(L)"           "PrN(R)"          
#> [45] "ProCN(L)"         "ProCN(R)"         "PDMN(R)"          "ProAN(L)"        
#> [49] "ProAN(R)"         "ProLN(L)"         "ProLN(R)"         "VProN(L)"        
#> [53] "VProN(R)"         "WTct(UTct-T2)(L)" "WTct(UTct-T2)(R)" "mVAC(T1)(L)"     
#> [57] "mVAC(T1)(R)"      "mVAC(T2)(L)"      "mVAC(T2)(R)"      "mVAC(T3)(L)"     
#> [61] "mVAC(T3)(R)"     
#> 
#> $`manc:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/manc-connectome"
#> 
#> $`manc:v1.0`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method     path                                          
#> [1,] "PROPFIND" "/api"                                        
#> [2,] "POST"     "/api/npexplorer/findneurons"                 
#> [3,] "GET"      "/api/cached/roicompleteness"                 
#> [4,] "GET"      "/api/v:ver/npexplorer/rankedtable"           
#> [5,] "GET"      "/api/dbmeta/version"                         
#> [6,] "POST"     "/api/v:ver/raw/cypher/transaction/:id/cypher"
#>      name                                                                                       
#> [1,] "github.com/labstack/echo/v4.glob..func1"                                                  
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getFindNeurons-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getROICompleteness-fm"   
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getVersion-fm"            
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }