Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "SPS(R)"            "EBr3d"             "SNP(L)"           
#>   [4] "FBl2"              "AVLP(R)"           "b2(R)"            
#>   [7] "MB(L)"             "BU(R)"             "ATL(R)"           
#>  [10] "a'3(R)"            "SMP(L)"            "EBr6"             
#>  [13] "bL(L)"             "a3(R)"             "OL(R)"            
#>  [16] "PB(L9)"            "LO(R)"             "FBl4"             
#>  [19] "a'L(L)"            "mALT(L)"           "PRW"              
#>  [22] "CA(R)"             "VLNP(R)"           "POC"              
#>  [25] "AMMC"              "EB"                "FBl3"             
#>  [28] "SMP(R)"            "GOR(L)"            "ME(R)"            
#>  [31] "b1(R)"             "IPS(R)"            "PB(R1)"           
#>  [34] "EPA(L)"            "bL(R)"             "NO(L)"            
#>  [37] "b'2(R)"            "SPS(L)"            "NO"               
#>  [40] "PB(R9)"            "gL(R)"             "g1(R)"            
#>  [43] "a'2(R)"            "LAL(L)"            "a2(R)"            
#>  [46] "PB(L6)"            "EBr3pw"            "PB(L7)"           
#>  [49] "RUB(L)"            "PENP"              "EBr1"             
#>  [52] "FLA(R)"            "AL(L)"             "ICL(L)"           
#>  [55] "dACA(R)"           "PB(L8)"            "a'1(R)"           
#>  [58] "PB(R6)"            "AOTU(R)"           "GC"               
#>  [61] "aL(R)"             "BU(L)"             "VES(R)"           
#>  [64] "PVLP(R)"           "PB(L4)"            "IB"               
#>  [67] "LOP(R)"            "CA(L)"             "NO3(R)"           
#>  [70] "LX(R)"             "EBr2r4"            "MB(+ACA)(R)"      
#>  [73] "aL(L)"             "AL(R)"             "GF(R)"            
#>  [76] "hemibrain"         "SCL(R)"            "PB(R8)"           
#>  [79] "CRE(L)"            "PB(L3)"            "ICL(R)"           
#>  [82] "CAN(R)"            "b'1(R)"            "vACA(R)"          
#>  [85] "LAL(-GA)(R)"       "lACA(R)"           "gL(L)"            
#>  [88] "NO(R)"             "FBl1"              "a1(R)"            
#>  [91] "INP"               "b'L(L)"            "PB(R7)"           
#>  [94] "VMNP"              "ROB(R)"            "FBl8"             
#>  [97] "PB(R2)"            "PB(R5)"            "ATL(L)"           
#> [100] "SAD"               "PED(R)"            "mALT(R)"          
#> [103] "g2(R)"             "LAL(R)"            "SIP(L)"           
#> [106] "WED(R)"            "MB(R)"             "NO2(R)"           
#> [109] "VES(L)"            "RUB(R)"            "GA(R)"            
#> [112] "SNP(R)"            "EPA(R)"            "NO3(L)"           
#> [115] "FBl6"              "PB"                "AOT(R)"           
#> [118] "EBr3am"            "SAD(-AMMC)"        "SLP(R)"           
#> [121] "AME(R)"            "PB(L2)"            "g3(R)"            
#> [124] "NO1(L)"            "FBl7"              "FBl9"             
#> [127] "PB(R3)"            "EBr5"              "PB(L1)"           
#> [130] "NO1(R)"            "LH(R)"             "LX(L)"            
#> [133] "GOR(R)"            "AB(L)"             "NO2(L)"           
#> [136] "b'L(R)"            "FBl5"              "AB(R)"            
#> [139] "PB(R4)"            "GNG"               "g4(R)"            
#> [142] "g5(R)"             "CX"                "CRE(R)"           
#> [145] "FB"                "a'L(R)"            "SCL(L)"           
#> [148] "CRE(-ROB,-RUB)(R)" "PLP(R)"            "SIP(R)"           
#> [151] "PB(L5)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method   path                                     
#> [1,] "POST"   "/api"                                   
#> [2,] "DELETE" "/api/*"                                 
#> [3,] "POST"   "/api/raw/cypher/transaction/:id/cypher" 
#> [4,] "POST"   "/api/raw/cypher/transaction/:id/kill"   
#> [5,] "GET"    "/api/npexplorer/simpleconnections"      
#> [6,] "GET"    "/api/v:ver/npexplorer/simpleconnections"
#>      name                                                                                             
#> [1,] "github.com/labstack/echo.glob..func1"                                                           
#> [2,] "github.com/labstack/echo.glob..func1"                                                           
#> [3,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                         
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                         
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }