Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0`
#> $`hemibrain:v1.0`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0`$ROIs
#>   [1] "CRE(-ROB,-RUB)(R)" "CA(L)"             "a2(R)"            
#>   [4] "SAD"               "NO3(R)"            "CRE(L)"           
#>   [7] "a'1(R)"            "SAD(-AMMC)"        "EBr5"             
#>  [10] "NO(R)"             "MB(R)"             "EPA(L)"           
#>  [13] "PED(R)"            "MB(L)"             "EB"               
#>  [16] "SIP(R)"            "NO"                "aL(L)"            
#>  [19] "g3(R)"             "g4(R)"             "AME(R)"           
#>  [22] "GNG"               "ATL(L)"            "FB"               
#>  [25] "SPS(L)"            "g5(R)"             "FBl4"             
#>  [28] "PB(R4)"            "FBl5"              "EBr1"             
#>  [31] "PB(R8)"            "FBl3"              "NO(L)"            
#>  [34] "SNP(L)"            "b2(R)"             "PB(R9)"           
#>  [37] "b1(R)"             "a3(R)"             "SMP(R)"           
#>  [40] "PB(R7)"            "EBr6"              "EBr3d"            
#>  [43] "a'L(R)"            "NO3(L)"            "FBl6"             
#>  [46] "GOR(L)"            "b'L(R)"            "LX(R)"            
#>  [49] "gL(L)"             "PB(L7)"            "a'2(R)"           
#>  [52] "LO(R)"             "a'L(L)"            "mALT(R)"          
#>  [55] "PB(R1)"            "CA(R)"             "ATL(R)"           
#>  [58] "AB(R)"             "SLP(R)"            "GC"               
#>  [61] "AL(L)"             "IPS(R)"            "LAL(R)"           
#>  [64] "AB(L)"             "FLA(R)"            "PVLP(R)"          
#>  [67] "RUB(L)"            "OL(R)"             "SCL(L)"           
#>  [70] "GA(R)"             "bL(L)"             "PB(R2)"           
#>  [73] "INP"               "AOT(R)"            "PB"               
#>  [76] "lACA(R)"           "PB(L8)"            "gL(R)"            
#>  [79] "FBl7"              "FBl8"              "GOR(R)"           
#>  [82] "AOTU(R)"           "SPS(R)"            "VMNP"             
#>  [85] "mALT(L)"           "SIP(L)"            "bL(R)"            
#>  [88] "VLNP(R)"           "RUB(R)"            "aL(R)"            
#>  [91] "PB(L9)"            "PB(R5)"            "IB"               
#>  [94] "EPA(R)"            "ROB(R)"            "FBl9"             
#>  [97] "b'2(R)"            "dACA(R)"           "BU(R)"            
#> [100] "LX(L)"             "EBr3am"            "CX"               
#> [103] "LOP(R)"            "PENP"              "GF(R)"            
#> [106] "LAL(-GA)(R)"       "CRE(R)"            "FBl1"             
#> [109] "PB(R3)"            "WED(R)"            "g1(R)"            
#> [112] "PB(L1)"            "ME(R)"             "a1(R)"            
#> [115] "NO2(L)"            "SMP(L)"            "MB(+ACA)(R)"      
#> [118] "LH(R)"             "POC"               "b'1(R)"           
#> [121] "AMMC"              "PB(L3)"            "PB(L6)"           
#> [124] "ICL(R)"            "CAN(R)"            "b'L(L)"           
#> [127] "a'3(R)"            "LAL(L)"            "BU(L)"            
#> [130] "AVLP(R)"           "NO2(R)"            "VES(L)"           
#> [133] "PB(R6)"            "SNP(R)"            "EBr3pw"           
#> [136] "FBl2"              "AL(R)"             "PB(L2)"           
#> [139] "g2(R)"             "NO1(R)"            "EBr2r4"           
#> [142] "ICL(L)"            "vACA(R)"           "PB(L5)"           
#> [145] "hemibrain"         "NO1(L)"            "VES(R)"           
#> [148] "PB(L4)"            "PRW"               "SCL(R)"           
#> [151] "PLP(R)"           
#> 
#> $`hemibrain:v1.0`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                     
#> [1,] "GET"  "/api/v:ver/cached/roiconnectivity"      
#> [2,] "GET"  "/api/npexplorer/neuronmetavals"         
#> [3,] "GET"  "/api/v:ver/npexplorer/simpleconnections"
#> [4,] "GET"  "/api/npexplorer/commonconnectivity"     
#> [5,] "POST" "/api/skeletons/skeleton/:dataset/:id"   
#> [6,] "GET"  "/*"                                     
#>      name                                                                                              
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getNeuronMetaVals-fm"    
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCommonConnectivity-fm"
#> [5,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                          
#> [6,] "github.com/labstack/echo.common.static.func1"                                                    

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }