Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "SCL(R)"            "SAD(-AMMC)"        "PVLP(R)"          
#>   [4] "PB(L1)"            "AL-VM7d(L)"        "b'1(R)"           
#>   [7] "AL-VC2(R)"         "dACA(R)"           "FLA(R)"           
#>  [10] "PB(R3)"            "LH(R)"             "SPS(L)"           
#>  [13] "ROB(R)"            "FBl9"              "EB"               
#>  [16] "gL(L)"             "PB(R1)"            "PB(L4)"           
#>  [19] "AL-VA2(R)"         "a'3(R)"            "CRE(-ROB,-RUB)(R)"
#>  [22] "EBr5"              "AL-DC4(L)"         "AL-DL5(L)"        
#>  [25] "AL-V(R)"           "g5(R)"             "AVLP(R)"          
#>  [28] "GOR(R)"            "VLNP(R)"           "NO2(L)"           
#>  [31] "GA(R)"             "PED(R)"            "mALT(R)"          
#>  [34] "AL-VM3(R)"         "b2(R)"             "AL-DM2(L)"        
#>  [37] "AL-DC4(R)"         "NO2(R)"            "AL-DL2d(R)"       
#>  [40] "AL-VA4(R)"         "PB(L2)"            "SNP(R)"           
#>  [43] "a'1(R)"            "AL-VA1d(R)"        "AL-VA5(R)"        
#>  [46] "IB"                "SIP(L)"            "lACA(R)"          
#>  [49] "g1(R)"             "AL-VM5d(R)"        "hemibrain"        
#>  [52] "PB(R7)"            "ME(R)"             "AL-VP2(R)"        
#>  [55] "PB(R4)"            "AL-DM3(R)"         "AL-DM6(R)"        
#>  [58] "LO(R)"             "IB112"             "PB(R5)"           
#>  [61] "NO(L)"             "AL-VL1(R)"         "gL(R)"            
#>  [64] "MB(+ACA)(R)"       "AB(R)"             "AL-DM3(L)"        
#>  [67] "AL-VM7v(L)"        "bL(L)"             "aL(R)"            
#>  [70] "MB(R)"             "AL-DM2(R)"         "EBr3am"           
#>  [73] "BU(R)"             "AL-VM5v(R)"        "AL-VP4(R)"        
#>  [76] "AL-VL2a(R)"        "NO(R)"             "ATL(R)"           
#>  [79] "ICL(R)"            "SMP(L)"            "VES(R)"           
#>  [82] "CAN(R)"            "FBl4"              "AL(L)"            
#>  [85] "AB(L)"             "AL-DA4m(R)"        "AL-VA7m(R)"       
#>  [88] "AL-VM7v(R)"        "MB(L)"             "FBl5"             
#>  [91] "LX(L)"             "IPS(R)"            "AL-VM7d(R)"       
#>  [94] "PRW"               "NO1(L)"            "g2(R)"            
#>  [97] "FBl8"              "aL(L)"             "AL-VP1l(R)"       
#> [100] "AL-DL1(R)"         "AL-DL2v(R)"        "PB(R6)"           
#> [103] "g4(R)"             "OL(R)"             "AL-VM1(R)"        
#> [106] "b'2(R)"            "AL-DM6(L)"         "a3(R)"            
#> [109] "FB-column3"        "AL-VA6(R)"         "ATL(L)"           
#> [112] "AL-DL3(R)"         "NO1(R)"            "PB(L6)"           
#> [115] "AL-DA1(R)"         "LOP(R)"            "a1(R)"            
#> [118] "EPA(R)"            "CRE(L)"            "bL(R)"            
#> [121] "AL-DM1(R)"         "RUB(R)"            "PENP"             
#> [124] "AL-DM4(L)"         "NO"                "LX(R)"            
#> [127] "AL-VC4(R)"         "AL-VA7l(R)"        "AOT(R)"           
#> [130] "SPS(R)"            "AL-DL4(L)"         "FBl1"             
#> [133] "SCL(L)"            "PLP(R)"            "EPA(L)"           
#> [136] "AL-DA4l(R)"        "AL-VL2p(R)"        "AL-DC2(L)"        
#> [139] "SIP(R)"            "AL-VC3l(R)"        "PB(L9)"           
#> [142] "AL-DM4(R)"         "FBl3"              "LAL(R)"           
#> [145] "AL-DC3(R)"         "GOR(L)"            "EBr1"             
#> [148] "AL-DC1(R)"         "AL-VP3(R)"         "EBr6"             
#> [151] "VMNP"              "EBr3d"             "AL-VC3m(R)"       
#> [154] "WED(R)"            "EBr2r4"            "AL-DM1(L)"        
#> [157] "AL-VM4(R)"         "SAD"               "b'L(R)"           
#> [160] "FBl2"              "CX"                "GF(R)"            
#> [163] "SNP(L)"            "AL-DA3(R)"         "mALT(L)"          
#> [166] "FB"                "RUB(L)"            "vACA(R)"          
#> [169] "AL-VP1m(R)"        "PB(L8)"            "PB(R9)"           
#> [172] "AL-D(R)"           "PB(L5)"            "GC"               
#> [175] "a'L(R)"            "VES(L)"            "AL-DA3(L)"        
#> [178] "PB(R2)"            "CA(L)"             "AL-VC1(R)"        
#> [181] "AL-VP5(R)"         "LAL(-GA)(R)"       "PB(L7)"           
#> [184] "AL-DA2(R)"         "ICL(L)"            "AL-DL4(R)"        
#> [187] "AL-DP1m(R)"        "a2(R)"             "BU(L)"            
#> [190] "AL-D(L)"           "FBl7"              "b1(R)"            
#> [193] "SLP(R)"            "g3(R)"             "PB"               
#> [196] "NO3(L)"            "AL-DA4m(L)"        "FBl6"             
#> [199] "AL-DP1l(R)"        "AMMC"              "POC"              
#> [202] "CRE(-RUB)(L)"      "b'L(L)"            "AL-DC2(R)"        
#> [205] "AL-DA2(L)"         "AL-VA6(L)"         "NO3(R)"           
#> [208] "CRE(R)"            "a'2(R)"            "AL-DC1(L)"        
#> [211] "INP"               "SMP(R)"            "a'L(L)"           
#> [214] "AL-VC5(R)"         "AL-VM2(R)"         "AL-VP1d(R)"       
#> [217] "LAL(L)"            "AOTU(R)"           "CA(R)"            
#> [220] "AL-DL5(R)"         "AL-DM5(R)"         "AL-VA1v(R)"       
#> [223] "EBr3pw"            "GNG"               "AL(R)"            
#> [226] "AL-DM5(L)"         "AL-DP1m(L)"        "AME(R)"           
#> [229] "PB(L3)"            "PB(R8)"            "AL-VA3(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(L9)"            "IB"                "SPS(L)"           
#>   [4] "PB(R1)"            "ICL(L)"            "CA(R)"            
#>   [7] "LH(R)"             "PB(R5)"            "EBr3am"           
#>  [10] "BU(L)"             "bL(L)"             "FBl7"             
#>  [13] "MB(L)"             "AMMC"              "CX"               
#>  [16] "CA(L)"             "PB(L3)"            "PB(L5)"           
#>  [19] "CRE(-ROB,-RUB)(R)" "PB(L4)"            "FBl4"             
#>  [22] "AB(R)"             "NO2(R)"            "VES(R)"           
#>  [25] "b1(R)"             "a3(R)"             "PB(L2)"           
#>  [28] "NO3(R)"            "NO(R)"             "a'1(R)"           
#>  [31] "AVLP(R)"           "LX(L)"             "aL(R)"            
#>  [34] "PVLP(R)"           "ATL(R)"            "CAN(R)"           
#>  [37] "RUB(R)"            "a2(R)"             "dACA(R)"          
#>  [40] "FBl1"              "WED(R)"            "AME(R)"           
#>  [43] "GNG"               "LAL(L)"            "SMP(L)"           
#>  [46] "SIP(L)"            "ROB(R)"            "PB(L7)"           
#>  [49] "NO(L)"             "a'L(L)"            "VES(L)"           
#>  [52] "SAD"               "CRE(R)"            "PB(L1)"           
#>  [55] "SNP(L)"            "PB(R2)"            "GF(R)"            
#>  [58] "b'L(R)"            "PB(L6)"            "PB(R7)"           
#>  [61] "EBr1"              "SNP(R)"            "vACA(R)"          
#>  [64] "CRE(L)"            "GOR(R)"            "b'1(R)"           
#>  [67] "a'3(R)"            "NO2(L)"            "GOR(L)"           
#>  [70] "aL(L)"             "AB(L)"             "SIP(R)"           
#>  [73] "PB(R4)"            "NO1(R)"            "VMNP"             
#>  [76] "g4(R)"             "FB"                "AOT(R)"           
#>  [79] "SAD(-AMMC)"        "FBl3"              "bL(R)"            
#>  [82] "PRW"               "SPS(R)"            "ME(R)"            
#>  [85] "NO1(L)"            "PB"                "EBr3d"            
#>  [88] "EB"                "AL(L)"             "FBl8"             
#>  [91] "g3(R)"             "PED(R)"            "g1(R)"            
#>  [94] "IPS(R)"            "MB(R)"             "AL(R)"            
#>  [97] "g2(R)"             "EBr6"              "EBr5"             
#> [100] "PB(R8)"            "SLP(R)"            "PENP"             
#> [103] "b'2(R)"            "AOTU(R)"           "mALT(L)"          
#> [106] "VLNP(R)"           "LO(R)"             "FBl5"             
#> [109] "RUB(L)"            "NO"                "PB(R9)"           
#> [112] "NO3(L)"            "LAL(R)"            "SMP(R)"           
#> [115] "lACA(R)"           "FLA(R)"            "b'L(L)"           
#> [118] "OL(R)"             "BU(R)"             "FBl2"             
#> [121] "INP"               "a'L(R)"            "gL(L)"            
#> [124] "EPA(R)"            "EBr3pw"            "MB(+ACA)(R)"      
#> [127] "GA(R)"             "PLP(R)"            "EBr2r4"           
#> [130] "a1(R)"             "POC"               "LX(R)"            
#> [133] "FBl9"              "EPA(L)"            "PB(L8)"           
#> [136] "g5(R)"             "LOP(R)"            "gL(R)"            
#> [139] "FBl6"              "b2(R)"             "a'2(R)"           
#> [142] "SCL(L)"            "ATL(L)"            "PB(R6)"           
#> [145] "LAL(-GA)(R)"       "mALT(R)"           "hemibrain"        
#> [148] "GC"                "PB(R3)"            "SCL(R)"           
#> [151] "ICL(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DL3(R)"         "BU(L)"             "AL-DM3(L)"        
#>   [4] "AL-VA1d(R)"        "AL-VM5v(R)"        "SCL(L)"           
#>   [7] "LAL(L)"            "FB"                "EBr3d"            
#>  [10] "b'2(R)"            "FLA(R)"            "AL(L)"            
#>  [13] "NO(L)"             "MB(R)"             "PLP(R)"           
#>  [16] "CAN(R)"            "AL-VM5d(R)"        "INP"              
#>  [19] "AL-DP1m(R)"        "PB(L8)"            "EPA(R)"           
#>  [22] "AL-DA4m(L)"        "AL-DM4(L)"         "FBl2"             
#>  [25] "AL-VC3m(R)"        "WED(R)"            "LO(R)"            
#>  [28] "PB(R5)"            "CRE(R)"            "VMNP"             
#>  [31] "PB(R2)"            "AMMC"              "AL-DP1l(R)"       
#>  [34] "AL-VA6(R)"         "AL-VP1l(R)"        "VLNP(R)"          
#>  [37] "FBl3"              "CA(R)"             "AL-DM6(R)"        
#>  [40] "LX(R)"             "AVLP(R)"           "AL-VP2(R)"        
#>  [43] "PENP"              "AOTU(R)"           "PB(L7)"           
#>  [46] "GOR(R)"            "AL-VA5(R)"         "CX"               
#>  [49] "SAD"               "PVLP(R)"           "PB(R4)"           
#>  [52] "EBr3am"            "CRE(-ROB,-RUB)(R)" "SMP(L)"           
#>  [55] "PB"                "a'2(R)"            "LH(R)"            
#>  [58] "AL-VC5(R)"         "AL-VM7v(R)"        "a'L(R)"           
#>  [61] "gL(L)"             "g5(R)"             "AL-VP1d(R)"       
#>  [64] "PED(R)"            "PB(L1)"            "AL-DA3(L)"        
#>  [67] "PB(L4)"            "AL-DC1(R)"         "AL-VP1m(R)"       
#>  [70] "AL-DM1(R)"         "AL(R)"             "AL-VM2(R)"        
#>  [73] "AL-VM7d(R)"        "ATL(L)"            "AL-DL5(R)"        
#>  [76] "EBr2r4"            "VES(R)"            "AL-VC2(R)"        
#>  [79] "AOT(R)"            "AL-V(R)"           "NO1(L)"           
#>  [82] "AL-VM7d(L)"        "ROB(R)"            "FB-column3"       
#>  [85] "FBl1"              "LX(L)"             "PB(L3)"           
#>  [88] "ME(R)"             "NO(R)"             "AL-DC3(R)"        
#>  [91] "IB"                "CRE(-RUB)(L)"      "bL(R)"            
#>  [94] "CA(L)"             "aL(L)"             "AL-DL2v(R)"       
#>  [97] "GOR(L)"            "g2(R)"             "NO"               
#> [100] "gL(R)"             "FBl8"              "ICL(R)"           
#> [103] "dACA(R)"           "AL-DL1(R)"         "AL-VC1(R)"        
#> [106] "b'1(R)"            "GA(R)"             "GNG"              
#> [109] "AL-DA4l(R)"        "AL-VA1v(R)"        "AL-VP5(R)"        
#> [112] "AL-DL4(L)"         "CRE(L)"            "AL-VM4(R)"        
#> [115] "PB(L5)"            "AL-DC1(L)"         "AME(R)"           
#> [118] "AL-VA4(R)"         "AL-VP4(R)"         "AL-DC2(L)"        
#> [121] "AL-D(L)"           "AL-DM4(R)"         "EBr1"             
#> [124] "AL-DP1m(L)"        "PB(R9)"            "AB(L)"            
#> [127] "RUB(R)"            "SIP(L)"            "aL(R)"            
#> [130] "NO1(R)"            "SLP(R)"            "AL-D(R)"          
#> [133] "g1(R)"             "AL-VL2p(R)"        "AL-DA4m(R)"       
#> [136] "EPA(L)"            "AL-DC4(R)"         "SMP(R)"           
#> [139] "GC"                "b'L(R)"            "a'3(R)"           
#> [142] "a'1(R)"            "ATL(R)"            "AL-DL4(R)"        
#> [145] "AL-VL2a(R)"        "BU(R)"             "AL-DL5(L)"        
#> [148] "SAD(-AMMC)"        "MB(+ACA)(R)"       "AL-VM1(R)"        
#> [151] "PB(R6)"            "SIP(R)"            "a'L(L)"           
#> [154] "AL-VL1(R)"         "AL-DA2(R)"         "PB(R8)"           
#> [157] "SNP(R)"            "AL-DM6(L)"         "AL-VM7v(L)"       
#> [160] "bL(L)"             "b'L(L)"            "AL-DA2(L)"        
#> [163] "AL-VA6(L)"         "AL-VA3(R)"         "AL-VA7l(R)"       
#> [166] "SCL(R)"            "FBl5"              "PB(R3)"           
#> [169] "AL-DA3(R)"         "NO2(R)"            "POC"              
#> [172] "AL-VC4(R)"         "LOP(R)"            "PRW"              
#> [175] "g4(R)"             "AL-DA1(R)"         "lACA(R)"          
#> [178] "PB(L9)"            "FBl6"              "PB(L2)"           
#> [181] "FBl9"              "LAL(R)"            "NO2(L)"           
#> [184] "hemibrain"         "EBr5"              "b1(R)"            
#> [187] "AL-DM2(R)"         "NO3(R)"            "PB(R1)"           
#> [190] "ICL(L)"            "FBl4"              "AB(R)"            
#> [193] "AL-VA7m(R)"        "MB(L)"             "SPS(L)"           
#> [196] "SNP(L)"            "AL-DM5(R)"         "AL-VC3l(R)"       
#> [199] "AL-VP3(R)"         "a2(R)"             "FBl7"             
#> [202] "AL-VM3(R)"         "mALT(L)"           "PB(R7)"           
#> [205] "AL-DM5(L)"         "AL-DM3(R)"         "mALT(R)"          
#> [208] "vACA(R)"           "LAL(-GA)(R)"       "RUB(L)"           
#> [211] "OL(R)"             "AL-VA2(R)"         "AL-DL2d(R)"       
#> [214] "g3(R)"             "AL-DM1(L)"         "AL-DM2(L)"        
#> [217] "PB(L6)"            "EBr3pw"            "a3(R)"            
#> [220] "IPS(R)"            "AL-DC2(R)"         "GF(R)"            
#> [223] "SPS(R)"            "b2(R)"             "EBr6"             
#> [226] "EB"                "AL-DC4(L)"         "a1(R)"            
#> [229] "NO3(L)"            "VES(L)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "gL(R)"             "PB(L4)"            "AL-VA5(R)"        
#>   [4] "PB(L3)"            "g2(R)"             "AL-VP3(R)"        
#>   [7] "AL-DP1m(L)"        "AL-VM7d(R)"        "IB"               
#>  [10] "AOTU(R)"           "ME(R)"             "FBl3"             
#>  [13] "dACA(R)"           "MB(+ACA)(R)"       "SLP(R)"           
#>  [16] "AL-DC4(R)"         "AL-DM3(L)"         "AL-VC4(R)"        
#>  [19] "a1(R)"             "LAL(L)"            "FB-column3"       
#>  [22] "AL-DA4l(R)"        "b2(R)"             "AL-DA3(L)"        
#>  [25] "ROB(R)"            "ICL(R)"            "AL-DC1(R)"        
#>  [28] "AL-DL5(R)"         "SIP(L)"            "a2(R)"            
#>  [31] "PED(R)"            "LAL(-GA)(R)"       "CRE(R)"           
#>  [34] "AL-VA6(R)"         "AL-VP1l(R)"        "VLNP(R)"          
#>  [37] "SMP(R)"            "GC"                "EBr5"             
#>  [40] "PB(L6)"            "AL-DL4(R)"         "mALT(L)"          
#>  [43] "PB(R9)"            "AL-VA4(R)"         "AL-VC3l(R)"       
#>  [46] "AL-VM5d(R)"        "CA(L)"             "EPA(R)"           
#>  [49] "INP"               "MB(R)"             "AL-VA3(R)"        
#>  [52] "AL-VC1(R)"         "a'3(R)"            "AL-DA4m(L)"       
#>  [55] "FBl2"              "AL-DM4(R)"         "PENP"             
#>  [58] "PB(R6)"            "EBr3d"             "NO"               
#>  [61] "AL-VM5v(R)"        "b'L(R)"            "EBr2r4"           
#>  [64] "CRE(L)"            "FBl4"              "AL-DM4(L)"        
#>  [67] "VES(L)"            "FBl7"              "NO2(R)"           
#>  [70] "AL-VC3m(R)"        "PB(L8)"            "SPS(L)"           
#>  [73] "AL-V(R)"           "GOR(R)"            "AL-VL2a(R)"       
#>  [76] "PB(L5)"            "b'2(R)"            "SNP(R)"           
#>  [79] "AME(R)"            "OL(R)"             "AL-DM2(L)"        
#>  [82] "ICL(L)"            "AL-VP1m(R)"        "PRW"              
#>  [85] "MB(L)"             "PB"                "AL-DL4(L)"        
#>  [88] "AL-VL1(R)"         "AL-VL2p(R)"        "SAD"              
#>  [91] "AL-D(L)"           "a3(R)"             "NO1(L)"           
#>  [94] "g4(R)"             "AL-VM7v(L)"        "VES(R)"           
#>  [97] "AL-DC3(R)"         "AL-VC5(R)"         "SPS(R)"           
#> [100] "BU(R)"             "FBl1"              "GOR(L)"           
#> [103] "AL-DM5(R)"         "AL(L)"             "FBl9"             
#> [106] "b1(R)"             "AL-DM3(R)"         "GF(R)"            
#> [109] "AL-DM6(L)"         "PB(L9)"            "GNG"              
#> [112] "LAL(R)"            "AL-DM2(R)"         "AL-VM7v(R)"       
#> [115] "AB(L)"             "EBr6"              "AL-VM1(R)"        
#> [118] "AOT(R)"            "FBl5"              "PB(R4)"           
#> [121] "BU(L)"             "FLA(R)"            "AL-DC4(L)"        
#> [124] "AL-DL5(L)"         "ATL(R)"            "AL-VA2(R)"        
#> [127] "AL-VM4(R)"         "CRE(-ROB,-RUB)(R)" "NO2(L)"           
#> [130] "PVLP(R)"           "bL(R)"             "AL-VM2(R)"        
#> [133] "AL-VP1d(R)"        "PB(L1)"            "AB(R)"            
#> [136] "CAN(R)"            "aL(L)"             "mALT(R)"          
#> [139] "WED(R)"            "PLP(R)"            "LH(R)"            
#> [142] "AL-VA1d(R)"        "AL-VC2(R)"         "g3(R)"            
#> [145] "g1(R)"             "LX(R)"             "AL-DA1(R)"        
#> [148] "AL-DL1(R)"         "EBr3am"            "hemibrain"        
#> [151] "lACA(R)"           "AL-VM7d(L)"        "LO(R)"            
#> [154] "PB(R2)"            "AL(R)"             "aL(R)"            
#> [157] "PB(R8)"            "AL-DA2(L)"         "AL-DL2v(R)"       
#> [160] "b'1(R)"            "SNP(L)"            "EPA(L)"           
#> [163] "AL-DM6(R)"         "AL-VA7l(R)"        "a'L(L)"           
#> [166] "vACA(R)"           "GA(R)"             "VMNP"             
#> [169] "AL-DP1m(R)"        "CX"                "NO(L)"            
#> [172] "AL-DM1(R)"         "AL-VP5(R)"         "a'L(R)"           
#> [175] "SIP(R)"            "FBl6"              "IPS(R)"           
#> [178] "EB"                "EBr3pw"            "AVLP(R)"          
#> [181] "a'1(R)"            "bL(L)"             "g5(R)"            
#> [184] "AL-DL2d(R)"        "NO3(L)"            "SCL(L)"           
#> [187] "AL-DM1(L)"         "AL-DM5(L)"         "POC"              
#> [190] "FB"                "SAD(-AMMC)"        "b'L(L)"           
#> [193] "AMMC"              "NO3(R)"            "EBr1"             
#> [196] "FBl8"              "AL-DL3(R)"         "ATL(L)"           
#> [199] "a'2(R)"            "PB(R1)"            "AL-DA4m(R)"       
#> [202] "CA(R)"             "AL-VA1v(R)"        "AL-VP4(R)"        
#> [205] "PB(R5)"            "RUB(R)"            "PB(L7)"           
#> [208] "RUB(L)"            "AL-DP1l(R)"        "SCL(R)"           
#> [211] "gL(L)"             "AL-DA3(R)"         "PB(L2)"           
#> [214] "SMP(L)"            "CRE(-RUB)(L)"      "AL-DC2(R)"        
#> [217] "AL-D(R)"           "AL-VP2(R)"         "PB(R7)"           
#> [220] "PB(R3)"            "LX(L)"             "AL-DA2(R)"        
#> [223] "LOP(R)"            "NO1(R)"            "AL-DC1(L)"        
#> [226] "AL-VA6(L)"         "AL-VA7m(R)"        "AL-DC2(L)"        
#> [229] "NO(R)"             "AL-VM3(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`manc:v1.0`
#> $`manc:v1.0`$`last-mod`
#> [1] "2023-05-31 23:54:01"
#> 
#> $`manc:v1.0`$uuid
#> [1] "59b37970bc7a4341b9a3a965a0d6b402"
#> 
#> $`manc:v1.0`$ROIs
#>  [1] "ADMN(L)"          "VProN(R)"         "Ov(R)"            "AbN3(R)"         
#>  [5] "IntTct"           "AbN4(L)"          "CvN(R)"           "DMetaN(L)"       
#>  [9] "ProCN(L)"         "ProCN(R)"         "LegNp(T3)(L)"     "NTct(UTct-T1)(R)"
#> [13] "MesoLN(L)"        "VProN(L)"         "mVAC(T2)(R)"      "LegNp(T1)(R)"    
#> [17] "ProAN(L)"         "GF(R)"            "AbN4(R)"          "AbNT"            
#> [21] "MetaLN(L)"        "ProAN(R)"         "ProLN(L)"         "PrN(R)"          
#> [25] "LTct"             "ANm"              "WTct(UTct-T2)(R)" "AbN1(L)"         
#> [29] "DProN(L)"         "ProLN(R)"         "mVAC(T1)(R)"      "GF(L)"           
#> [33] "PDMN(L)"          "CvN(L)"           "PDMN(R)"          "ADMN(R)"         
#> [37] "mVAC(T1)(L)"      "mVAC(T2)(L)"      "Ov(L)"            "AbN1(R)"         
#> [41] "AbN2(L)"          "CV"               "mVAC(T3)(R)"      "AbN2(R)"         
#> [45] "DProN(R)"         "MesoAN(R)"        "LegNp(T2)(R)"     "NTct(UTct-T1)(L)"
#> [49] "HTct(UTct-T3)(L)" "HTct(UTct-T3)(R)" "MesoAN(L)"        "MesoLN(R)"       
#> [53] "LegNp(T3)(R)"     "PrN(L)"           "mVAC(T3)(L)"      "MetaLN(R)"       
#> [57] "LegNp(T2)(L)"     "WTct(UTct-T2)(L)" "AbN3(L)"          "DMetaN(R)"       
#> [61] "LegNp(T1)(L)"    
#> 
#> $`manc:v1.0`$superLevelROIs
#>  [1] "ADMN(L)"          "ADMN(R)"          "Ov(L)"            "Ov(R)"           
#>  [5] "ANm"              "AbN1(L)"          "AbN1(R)"          "AbN2(L)"         
#>  [9] "AbN2(R)"          "AbN3(L)"          "AbN3(R)"          "AbN4(L)"         
#> [13] "AbN4(R)"          "AbNT"             "CV"               "CvN(L)"          
#> [17] "CvN(R)"           "DMetaN(L)"        "DMetaN(R)"        "DProN(L)"        
#> [21] "DProN(R)"         "GF(L)"            "GF(R)"            "HTct(UTct-T3)(L)"
#> [25] "HTct(UTct-T3)(R)" "LegNp(T1)(L)"     "LegNp(T1)(R)"     "LegNp(T2)(L)"    
#> [29] "LegNp(T2)(R)"     "LegNp(T3)(L)"     "LegNp(T3)(R)"     "IntTct"          
#> [33] "LTct"             "MesoAN(L)"        "MesoAN(R)"        "MesoLN(L)"       
#> [37] "MesoLN(R)"        "MetaLN(L)"        "MetaLN(R)"        "NTct(UTct-T1)(L)"
#> [41] "NTct(UTct-T1)(R)" "PDMN(L)"          "PrN(L)"           "PrN(R)"          
#> [45] "ProCN(L)"         "ProCN(R)"         "PDMN(R)"          "ProAN(L)"        
#> [49] "ProAN(R)"         "ProLN(L)"         "ProLN(R)"         "VProN(L)"        
#> [53] "VProN(R)"         "WTct(UTct-T2)(L)" "WTct(UTct-T2)(R)" "mVAC(T1)(L)"     
#> [57] "mVAC(T1)(R)"      "mVAC(T2)(L)"      "mVAC(T2)(R)"      "mVAC(T3)(L)"     
#> [61] "mVAC(T3)(R)"     
#> 
#> $`manc:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/manc-connectome"
#> 
#> $`manc:v1.0`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method    path                                           
#> [1,] "POST"    "/api/v:ver/raw/cypher/transaction"            
#> [2,] "GET"     "/api/v:ver/npexplorer/celltype/:dataset/:type"
#> [3,] "CONNECT" "/api"                                         
#> [4,] "GET"     "/api/npexplorer/rankedtable"                  
#> [5,] "GET"     "/api/v:ver/npexplorer/simpleconnections"      
#> [6,] "POST"    "/api/v:ver/npexplorer/commonconnectivity"     
#>      name                                                                                       
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [3,] "github.com/labstack/echo/v4.glob..func1"                                                  
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getRankedTable-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }