Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DA2(L)"         "AL-DM1(L)"         "FBl9"             
#>   [4] "b'L(L)"            "AL-VP1m(R)"        "SCL(R)"           
#>   [7] "SIP(R)"            "ROB(R)"            "SNP(L)"           
#>  [10] "MB(+ACA)(R)"       "AL-VM7v(R)"        "SIP(L)"           
#>  [13] "g1(R)"             "AL-DA1(R)"         "AL-VA1v(R)"       
#>  [16] "a'2(R)"            "SNP(R)"            "PVLP(R)"          
#>  [19] "AL(R)"             "b'1(R)"            "AMMC"             
#>  [22] "AOTU(R)"           "SMP(R)"            "PB(L1)"           
#>  [25] "PB(R9)"            "bL(R)"             "VLNP(R)"          
#>  [28] "AL(L)"             "ICL(L)"            "AL-VM7d(R)"       
#>  [31] "POC"               "AL-DM3(L)"         "PB(R4)"           
#>  [34] "GOR(R)"            "AL-DP1m(R)"        "NO3(R)"           
#>  [37] "PB(R8)"            "g4(R)"             "NO2(R)"           
#>  [40] "AL-DM2(R)"         "AL-VA3(R)"         "AL-VP5(R)"        
#>  [43] "AL-DM6(L)"         "PB(R1)"            "NO(R)"            
#>  [46] "AL-DL2v(R)"        "AL-DP1l(R)"        "AL-VA7l(R)"       
#>  [49] "CA(L)"             "FBl8"              "b1(R)"            
#>  [52] "AL-DC2(R)"         "AL-D(R)"           "vACA(R)"          
#>  [55] "EBr6"              "AL-DL4(L)"         "FB"               
#>  [58] "LX(R)"             "INP"               "PB(L3)"           
#>  [61] "CRE(-RUB)(L)"      "PB(R3)"            "AL-VA5(R)"        
#>  [64] "PENP"              "FBl2"              "GNG"              
#>  [67] "AL-VL2a(R)"        "EBr1"              "AL-DC1(L)"        
#>  [70] "FBl1"              "AL-VL1(R)"         "SPS(R)"           
#>  [73] "PB(L8)"            "AL-DC2(L)"         "AL-DA4l(R)"       
#>  [76] "AL-DL3(R)"         "AL-V(R)"           "AL-DC1(R)"        
#>  [79] "AL-DL1(R)"         "a'3(R)"            "CX"               
#>  [82] "VES(R)"            "AL-VA2(R)"         "AL-VC4(R)"        
#>  [85] "AL-VM1(R)"         "AL-VM5d(R)"        "PB(L5)"           
#>  [88] "GF(R)"             "VMNP"              "SLP(R)"           
#>  [91] "AL-VA6(R)"         "GC"                "AL-DM5(L)"        
#>  [94] "EPA(L)"            "AL-DL4(R)"         "AB(L)"            
#>  [97] "ATL(L)"            "SCL(L)"            "a1(R)"            
#> [100] "CRE(R)"            "b2(R)"             "aL(R)"            
#> [103] "IB"                "bL(L)"             "EBr3pw"           
#> [106] "AL-VP2(R)"         "mALT(R)"           "gL(L)"            
#> [109] "AL-VM4(R)"         "AL-VM7d(L)"        "CA(R)"            
#> [112] "AL-VM2(R)"         "EPA(R)"            "g5(R)"            
#> [115] "LX(L)"             "LOP(R)"            "mALT(L)"          
#> [118] "MB(L)"             "FLA(R)"            "FBl5"             
#> [121] "AL-DP1m(L)"        "aL(L)"             "VES(L)"           
#> [124] "LO(R)"             "AL-VP4(R)"         "LAL(-GA)(R)"      
#> [127] "IB112"             "AL-DA2(R)"         "RUB(R)"           
#> [130] "AL-VL2p(R)"        "FBl6"              "AL-VM3(R)"        
#> [133] "NO1(R)"            "a'L(R)"            "g3(R)"            
#> [136] "NO(L)"             "AL-DM2(L)"         "IPS(R)"           
#> [139] "AL-DM5(R)"         "AL-DM6(R)"         "SMP(L)"           
#> [142] "SPS(L)"            "EBr3d"             "FBl7"             
#> [145] "CAN(R)"            "LAL(R)"            "AL-DA4m(R)"       
#> [148] "AL-VA7m(R)"        "CRE(-ROB,-RUB)(R)" "gL(R)"            
#> [151] "LAL(L)"            "PB"                "AL-DL5(L)"        
#> [154] "PB(R2)"            "ICL(R)"            "AL-VP1l(R)"       
#> [157] "PLP(R)"            "SAD(-AMMC)"        "AB(R)"            
#> [160] "PB(L4)"            "PB(L2)"            "GA(R)"            
#> [163] "b'L(R)"            "AL-DC4(R)"         "AL-VC1(R)"        
#> [166] "AL-VC5(R)"         "PB(R6)"            "PB(L7)"           
#> [169] "AL-DA3(R)"         "AL-DL5(R)"         "AL-VA4(R)"        
#> [172] "PRW"               "WED(R)"            "PB(L6)"           
#> [175] "FB-column3"        "a3(R)"             "dACA(R)"          
#> [178] "BU(R)"             "LH(R)"             "GOR(L)"           
#> [181] "EB"                "AL-D(L)"           "AL-VA1d(R)"       
#> [184] "SAD"               "AL-DM3(R)"         "PB(R5)"           
#> [187] "b'2(R)"            "a2(R)"             "AL-DA4m(L)"       
#> [190] "OL(R)"             "AL-VC2(R)"         "AL-VM5v(R)"       
#> [193] "AL-VM7v(L)"        "PB(R7)"            "PED(R)"           
#> [196] "NO1(L)"            "AL-DA3(L)"         "AL-VA6(L)"        
#> [199] "AL-VP3(R)"         "NO2(L)"            "EBr2r4"           
#> [202] "AME(R)"            "AL-DM1(R)"         "ME(R)"            
#> [205] "a'L(L)"            "AL-VC3m(R)"        "hemibrain"        
#> [208] "PB(L9)"            "AVLP(R)"           "AOT(R)"           
#> [211] "EBr5"              "AL-DC3(R)"         "CRE(L)"           
#> [214] "a'1(R)"            "AL-DC4(L)"         "NO3(L)"           
#> [217] "RUB(L)"            "MB(R)"             "lACA(R)"          
#> [220] "g2(R)"             "BU(L)"             "AL-DM4(L)"        
#> [223] "NO"                "FBl3"              "FBl4"             
#> [226] "AL-DM4(R)"         "AL-VP1d(R)"        "ATL(R)"           
#> [229] "AL-DL2d(R)"        "AL-VC3l(R)"        "EBr3am"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "FBl7"              "vACA(R)"           "GA(R)"            
#>   [4] "FLA(R)"            "PB(R3)"            "LX(L)"            
#>   [7] "GOR(R)"            "SMP(R)"            "PB(L8)"           
#>  [10] "gL(L)"             "PB(L4)"            "PB(L6)"           
#>  [13] "VES(R)"            "INP"               "ME(R)"            
#>  [16] "a'L(R)"            "MB(+ACA)(R)"       "NO(L)"            
#>  [19] "RUB(R)"            "LOP(R)"            "a'3(R)"           
#>  [22] "mALT(R)"           "PLP(R)"            "NO"               
#>  [25] "AB(R)"             "aL(L)"             "VES(L)"           
#>  [28] "PED(R)"            "CRE(R)"            "ATL(R)"           
#>  [31] "PB(L2)"            "ATL(L)"            "BU(L)"            
#>  [34] "FBl3"              "FBl4"              "PB(R6)"           
#>  [37] "SPS(R)"            "PVLP(R)"           "NO1(R)"           
#>  [40] "LX(R)"             "g2(R)"             "CRE(L)"           
#>  [43] "b1(R)"             "AME(R)"            "RUB(L)"           
#>  [46] "b'L(L)"            "PB(L3)"            "SIP(L)"           
#>  [49] "LAL(R)"            "PB(R5)"            "VMNP"             
#>  [52] "PB(L7)"            "PRW"               "PB(R8)"           
#>  [55] "EBr2r4"            "POC"               "SCL(L)"           
#>  [58] "LAL(L)"            "g3(R)"             "PB(R4)"           
#>  [61] "IB"                "AMMC"              "EB"               
#>  [64] "VLNP(R)"           "FBl6"              "NO1(L)"           
#>  [67] "BU(R)"             "FBl5"              "GNG"              
#>  [70] "a2(R)"             "aL(R)"             "PB(R9)"           
#>  [73] "mALT(L)"           "gL(R)"             "ICL(L)"           
#>  [76] "g5(R)"             "LH(R)"             "LAL(-GA)(R)"      
#>  [79] "AL(L)"             "g1(R)"             "SLP(R)"           
#>  [82] "SMP(L)"            "hemibrain"         "a'L(L)"           
#>  [85] "NO2(L)"            "a'2(R)"            "bL(L)"            
#>  [88] "MB(R)"             "SPS(L)"            "FBl2"             
#>  [91] "GC"                "AVLP(R)"           "CAN(R)"           
#>  [94] "bL(R)"             "OL(R)"             "lACA(R)"          
#>  [97] "PB(R7)"            "PB(R1)"            "b'1(R)"           
#> [100] "EBr3am"            "CA(L)"             "PB(L1)"           
#> [103] "SNP(R)"            "CA(R)"             "PB(L5)"           
#> [106] "WED(R)"            "EBr6"              "a3(R)"            
#> [109] "AB(L)"             "a1(R)"             "EBr3pw"           
#> [112] "FBl8"              "NO(R)"             "IPS(R)"           
#> [115] "MB(L)"             "b'L(R)"            "CX"               
#> [118] "dACA(R)"           "EPA(R)"            "b2(R)"            
#> [121] "FB"                "EBr3d"             "g4(R)"            
#> [124] "a'1(R)"            "FBl9"              "AOTU(R)"          
#> [127] "NO3(R)"            "PB(L9)"            "SAD(-AMMC)"       
#> [130] "ICL(R)"            "NO2(R)"            "EPA(L)"           
#> [133] "EBr5"              "PB"                "GOR(L)"           
#> [136] "EBr1"              "FBl1"              "AL(R)"            
#> [139] "AOT(R)"            "PENP"              "GF(R)"            
#> [142] "SIP(R)"            "PB(R2)"            "CRE(-ROB,-RUB)(R)"
#> [145] "b'2(R)"            "SAD"               "NO3(L)"           
#> [148] "SNP(L)"            "LO(R)"             "SCL(R)"           
#> [151] "ROB(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "SMP(L)"            "g2(R)"             "SCL(R)"           
#>   [4] "VMNP"              "a2(R)"             "hemibrain"        
#>   [7] "CRE(R)"            "AL-VA1v(R)"        "AL-VP1l(R)"       
#>  [10] "AL-DA4m(L)"        "PB(L9)"            "GNG"              
#>  [13] "AL-DA2(L)"         "AL-DM1(L)"         "FBl8"             
#>  [16] "AL-VA4(R)"         "GA(R)"             "GC"               
#>  [19] "PB(L1)"            "AL-VA5(R)"         "AL-VP5(R)"        
#>  [22] "EBr2r4"            "VES(L)"            "AL-D(R)"          
#>  [25] "EB"                "NO(R)"             "FBl1"             
#>  [28] "LH(R)"             "AL-VC2(R)"         "mALT(R)"          
#>  [31] "MB(L)"             "EBr1"              "AL-VC3l(R)"       
#>  [34] "EBr5"              "AL-DL2v(R)"        "AL-VM7v(R)"       
#>  [37] "PB(R1)"            "g5(R)"             "FBl4"             
#>  [40] "AL-DL2d(R)"        "PB(L8)"            "CRE(-RUB)(L)"     
#>  [43] "NO3(L)"            "AL-DA3(R)"         "AL-DL5(R)"        
#>  [46] "CX"                "EBr3d"             "AL-VM7d(L)"       
#>  [49] "NO"                "ROB(R)"            "ATL(L)"           
#>  [52] "PB(R9)"            "SNP(L)"            "AL-VM2(R)"        
#>  [55] "PB(R4)"            "IB"                "SIP(R)"           
#>  [58] "ATL(R)"            "NO2(R)"            "aL(R)"            
#>  [61] "FBl6"              "AL-VA1d(R)"        "AL-VL2p(R)"       
#>  [64] "AL-DA4m(R)"        "AL-VC4(R)"         "AL-VM5v(R)"       
#>  [67] "BU(R)"             "AB(R)"             "GF(R)"            
#>  [70] "OL(R)"             "AL-VP3(R)"         "AL-VL2a(R)"       
#>  [73] "LX(L)"             "RUB(L)"            "AL-DA4l(R)"       
#>  [76] "FBl3"              "AL-DC3(R)"         "AL-VA3(R)"        
#>  [79] "SMP(R)"            "PB(L3)"            "AL-DL4(L)"        
#>  [82] "AL-DM3(L)"         "EPA(L)"            "LAL(R)"           
#>  [85] "AL-DA1(R)"         "b'2(R)"            "LX(R)"            
#>  [88] "AL-D(L)"           "AL-V(R)"           "LAL(L)"           
#>  [91] "g3(R)"             "a3(R)"             "AL-DC2(R)"        
#>  [94] "PED(R)"            "gL(L)"             "AL-DL5(L)"        
#>  [97] "AL-DL3(R)"         "CA(L)"             "NO2(L)"           
#> [100] "a'2(R)"            "AL-DA3(L)"         "MB(R)"            
#> [103] "AL-DC4(R)"         "AOT(R)"            "b'1(R)"           
#> [106] "dACA(R)"           "AL-DP1m(L)"        "LO(R)"            
#> [109] "EBr3pw"            "RUB(R)"            "bL(L)"            
#> [112] "FBl2"              "LAL(-GA)(R)"       "SPS(L)"           
#> [115] "g4(R)"             "AL-DM4(L)"         "PVLP(R)"          
#> [118] "AVLP(R)"           "EBr3am"            "PENP"             
#> [121] "SIP(L)"            "PB(R7)"            "a'3(R)"           
#> [124] "WED(R)"            "FB-column3"        "SCL(L)"           
#> [127] "AL-DM6(L)"         "a'1(R)"            "AL-DL1(R)"        
#> [130] "SAD"               "lACA(R)"           "FB"               
#> [133] "b'L(R)"            "AL-VM3(R)"         "AL-DL4(R)"        
#> [136] "AL-VC1(R)"         "AMMC"              "BU(L)"            
#> [139] "MB(+ACA)(R)"       "AL-VC3m(R)"        "FBl7"             
#> [142] "PRW"               "AL-DC2(L)"         "AL-DM1(R)"        
#> [145] "AL-VP4(R)"         "PB(R6)"            "CAN(R)"           
#> [148] "AL-VM7d(R)"        "AL-VM7v(L)"        "aL(L)"            
#> [151] "AL-VA7l(R)"        "SPS(R)"            "GOR(L)"           
#> [154] "AL-DC1(L)"         "GOR(R)"            "AL-VP1m(R)"       
#> [157] "PB(L5)"            "AOTU(R)"           "VES(R)"           
#> [160] "AME(R)"            "AL-DP1m(R)"        "AL-VC5(R)"        
#> [163] "ICL(L)"            "AL-VA2(R)"         "AL-VA7m(R)"       
#> [166] "IPS(R)"            "LOP(R)"            "bL(R)"            
#> [169] "b'L(L)"            "AL-VM1(R)"         "PB(L2)"           
#> [172] "PB(L7)"            "FBl5"              "PB(R3)"           
#> [175] "AL-VM4(R)"         "CRE(L)"            "VLNP(R)"          
#> [178] "AL-DM3(R)"         "PB(L4)"            "SLP(R)"           
#> [181] "AL(R)"             "ICL(R)"            "vACA(R)"          
#> [184] "mALT(L)"           "NO1(R)"            "a'L(R)"           
#> [187] "PB(R2)"            "CA(R)"             "AL-DM4(R)"        
#> [190] "FBl9"              "AL-VP1d(R)"        "b2(R)"            
#> [193] "PB(R5)"            "NO3(R)"            "INP"              
#> [196] "AL-VM5d(R)"        "a'L(L)"            "AL-DM5(R)"        
#> [199] "AL-VL1(R)"         "AL-VP2(R)"         "a1(R)"            
#> [202] "NO1(L)"            "FLA(R)"            "SAD(-AMMC)"       
#> [205] "NO(L)"             "AB(L)"             "AL-DM2(L)"        
#> [208] "gL(R)"             "EPA(R)"            "AL-DM5(L)"        
#> [211] "AL-DM2(R)"         "AL-VA6(R)"         "CRE(-ROB,-RUB)(R)"
#> [214] "PB"                "AL(L)"             "b1(R)"            
#> [217] "SNP(R)"            "AL-VA6(L)"         "AL-DC1(R)"        
#> [220] "AL-DP1l(R)"        "g1(R)"             "EBr6"             
#> [223] "ME(R)"             "AL-DM6(R)"         "PB(R8)"           
#> [226] "PB(L6)"            "AL-DA2(R)"         "PLP(R)"           
#> [229] "AL-DC4(L)"         "POC"              
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "gL(R)"             "CAN(R)"            "AL-DM2(R)"        
#>   [4] "AL-VC3m(R)"        "PB(L5)"            "VMNP"             
#>   [7] "SAD"               "a3(R)"             "AL-DA4m(R)"       
#>  [10] "bL(R)"             "AL-VA6(L)"         "FBl6"             
#>  [13] "POC"               "mALT(L)"           "EPA(R)"           
#>  [16] "LAL(L)"            "AL-VM3(R)"         "LOP(R)"           
#>  [19] "PB(R2)"            "WED(R)"            "PB(L1)"           
#>  [22] "AL-D(L)"           "FBl4"              "AL-DP1l(R)"       
#>  [25] "IB"                "SIP(L)"            "mALT(R)"          
#>  [28] "AME(R)"            "EPA(L)"            "AL-VA7m(R)"       
#>  [31] "GA(R)"             "PB(R8)"            "NO3(L)"           
#>  [34] "AL-DM5(R)"         "PRW"               "NO(L)"            
#>  [37] "AL(R)"             "EBr1"              "b'L(L)"           
#>  [40] "AL-VA5(R)"         "PB(L3)"            "CA(R)"            
#>  [43] "PB(R4)"            "AL-VL2a(R)"        "AL-V(R)"          
#>  [46] "PB(R5)"            "g1(R)"             "AL-DC4(L)"        
#>  [49] "SNP(L)"            "FBl1"              "SMP(R)"           
#>  [52] "AMMC"              "MB(L)"             "ATL(R)"           
#>  [55] "AL-VM7d(R)"        "SCL(R)"            "a'3(R)"           
#>  [58] "EBr3d"             "AL-DC2(L)"         "IPS(R)"           
#>  [61] "MB(R)"             "AB(L)"             "AL(L)"            
#>  [64] "ME(R)"             "aL(R)"             "b1(R)"            
#>  [67] "AL-VM2(R)"         "LX(L)"             "NO"               
#>  [70] "AL-VP4(R)"         "gL(L)"             "PB(R3)"           
#>  [73] "AL-DM3(L)"         "AL-VM4(R)"         "EBr5"             
#>  [76] "AL-VM1(R)"         "AL-VP1l(R)"        "a1(R)"            
#>  [79] "dACA(R)"           "b'L(R)"            "FLA(R)"           
#>  [82] "NO(R)"             "PB(L9)"            "FBl9"             
#>  [85] "VES(L)"            "SLP(R)"            "PB(R7)"           
#>  [88] "AL-DM4(R)"         "AL-DA4l(R)"        "ROB(R)"           
#>  [91] "AL-VC5(R)"         "SCL(L)"            "ATL(L)"           
#>  [94] "a'L(R)"            "NO3(R)"            "ICL(L)"           
#>  [97] "FBl2"              "ICL(R)"            "AL-DC2(R)"        
#> [100] "AL-DP1m(R)"        "PED(R)"            "AL-DA2(L)"        
#> [103] "FBl5"              "GOR(R)"            "a'L(L)"           
#> [106] "AL-VL1(R)"         "AL-VL2p(R)"        "AOTU(R)"          
#> [109] "AL-DL1(R)"         "AL-D(R)"           "PB(R9)"           
#> [112] "PLP(R)"            "OL(R)"             "g2(R)"            
#> [115] "SPS(L)"            "AL-DL3(R)"         "AL-VM7d(L)"       
#> [118] "EBr6"              "LX(R)"             "AL-DP1m(L)"       
#> [121] "AB(R)"             "SPS(R)"            "NO1(R)"           
#> [124] "CA(L)"             "CRE(R)"            "AL-DL4(L)"        
#> [127] "AL-DA1(R)"         "AL-DA2(R)"         "AL-VC1(R)"        
#> [130] "AL-VM7v(R)"        "RUB(L)"            "g3(R)"            
#> [133] "a'1(R)"            "AL-VA3(R)"         "RUB(R)"           
#> [136] "GF(R)"             "CRE(L)"            "SAD(-AMMC)"       
#> [139] "VES(R)"            "AL-DL2d(R)"        "AL-VP2(R)"        
#> [142] "AL-DM2(L)"         "LO(R)"             "AL-DA3(R)"        
#> [145] "AL-DC1(R)"         "AL-DM6(L)"         "MB(+ACA)(R)"      
#> [148] "GNG"               "BU(R)"             "LH(R)"            
#> [151] "SMP(L)"            "AL-VC3l(R)"        "GOR(L)"           
#> [154] "vACA(R)"           "PB(L8)"            "AL-DM1(L)"        
#> [157] "PVLP(R)"           "AL-DC3(R)"         "AL-VA1v(R)"       
#> [160] "b'1(R)"            "LAL(-GA)(R)"       "g5(R)"            
#> [163] "FBl8"              "AL-VA4(R)"         "SNP(R)"           
#> [166] "FB"                "AL-DC4(R)"         "AL-DM1(R)"        
#> [169] "AL-VM5d(R)"        "hemibrain"         "AL-DM4(L)"        
#> [172] "AL-DM5(L)"         "AL-DM6(R)"         "AL-VA2(R)"        
#> [175] "AL-VM5v(R)"        "BU(L)"             "AL-VA7l(R)"       
#> [178] "AL-VC2(R)"         "NO1(L)"            "CRE(-RUB)(L)"     
#> [181] "AL-DL2v(R)"        "EBr2r4"            "AL-DL5(L)"        
#> [184] "AL-DM3(R)"         "PB(R6)"            "lACA(R)"          
#> [187] "aL(L)"             "AOT(R)"            "CX"               
#> [190] "PENP"              "PB(L6)"            "bL(L)"            
#> [193] "LAL(R)"            "AL-VM7v(L)"        "FBl7"             
#> [196] "PB(L4)"            "SIP(R)"            "g4(R)"            
#> [199] "FB-column3"        "PB"                "FBl3"             
#> [202] "AL-VA6(R)"         "GC"                "a'2(R)"           
#> [205] "EBr3am"            "NO2(L)"            "b'2(R)"           
#> [208] "AL-DA3(L)"         "NO2(R)"            "AL-VA1d(R)"       
#> [211] "AL-VP1m(R)"        "EB"                "b2(R)"            
#> [214] "INP"               "PB(L2)"            "a2(R)"            
#> [217] "AL-DL5(R)"         "AL-VP5(R)"         "AL-VP3(R)"        
#> [220] "AL-DA4m(L)"        "PB(L7)"            "AL-DL4(R)"        
#> [223] "AL-DC1(L)"         "AVLP(R)"           "AL-VC4(R)"        
#> [226] "EBr3pw"            "VLNP(R)"           "CRE(-ROB,-RUB)(R)"
#> [229] "AL-VP1d(R)"        "PB(R1)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                               
#> [1,] "POST"    "/api/custom/custom"               
#> [2,] "CONNECT" "/api"                             
#> [3,] "POST"    "/api/npexplorer/simpleconnections"
#> [4,] "POST"    "/api/v:ver/raw/cypher/transaction"
#> [5,] "GET"     "/api/v:ver/dbmeta/datasets"       
#> [6,] "GET"     "/profile"                         
#>      name                                                                                             
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/custom.cypherAPI.getCustom-fm"                  
#> [2,] "github.com/labstack/echo.glob..func1"                                                           
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [6,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                         

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }