Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-DM1(R)"         "PB(L3)"            "lACA(R)"          
#>   [4] "NO1(L)"            "PB(L6)"            "AL-DM6(R)"        
#>   [7] "EBr1"              "AL-VM7v(L)"        "AL-DM5(L)"        
#>  [10] "BU(L)"             "VMNP"              "FBl3"             
#>  [13] "b'L(L)"            "SMP(L)"            "a'1(R)"           
#>  [16] "vACA(R)"           "RUB(L)"            "AL-DA2(R)"        
#>  [19] "AL-DM5(R)"         "AL-VM1(R)"         "IB"               
#>  [22] "bL(L)"             "AL-DC1(R)"         "AL-VA2(R)"        
#>  [25] "AL-VA6(L)"         "NO3(L)"            "AL-DC4(R)"        
#>  [28] "AL-VM5d(R)"        "NO2(L)"            "EBr3d"            
#>  [31] "EBr2r4"            "FBl9"              "AL-V(R)"          
#>  [34] "AB(L)"             "PED(R)"            "PB(R7)"           
#>  [37] "MB(+ACA)(R)"       "AB(R)"             "AL-DA3(R)"        
#>  [40] "ROB(R)"            "AL-VP1m(R)"        "PB(R5)"           
#>  [43] "BU(R)"             "AL-VA5(R)"         "SCL(L)"           
#>  [46] "AOT(R)"            "GF(R)"             "PB"               
#>  [49] "AL-DM3(L)"         "PB(L4)"            "AL-VM3(R)"        
#>  [52] "g3(R)"             "AL(L)"             "NO(R)"            
#>  [55] "PB(L7)"            "a3(R)"             "EB"               
#>  [58] "FBl2"              "PB(R4)"            "AL-DA4l(R)"       
#>  [61] "AL-VC1(R)"         "a'3(R)"            "SNP(R)"           
#>  [64] "SLP(R)"            "AL(R)"             "RUB(R)"           
#>  [67] "FBl1"              "AL-DL1(R)"         "AL-DP1l(R)"       
#>  [70] "PB(R3)"            "OL(R)"             "AL-DM3(R)"        
#>  [73] "b'2(R)"            "EBr6"              "EPA(R)"           
#>  [76] "AL-DC2(L)"         "AL-DL5(L)"         "b2(R)"            
#>  [79] "bL(R)"             "AL-DA4m(R)"        "PB(L1)"           
#>  [82] "AL-VC4(R)"         "PB(L2)"            "LH(R)"            
#>  [85] "AL-VC3m(R)"        "AL-DA1(R)"         "CX"               
#>  [88] "CRE(R)"            "CA(L)"             "GOR(L)"           
#>  [91] "g2(R)"             "PB(R1)"            "AL-VM5v(R)"       
#>  [94] "LOP(R)"            "SPS(R)"            "SCL(R)"           
#>  [97] "ICL(R)"            "SAD"               "FBl8"             
#> [100] "NO(L)"             "AL-VA3(R)"         "AL-VM7v(R)"       
#> [103] "AL-VP3(R)"         "GC"                "ME(R)"            
#> [106] "AL-DM2(L)"         "CAN(R)"            "AL-DL3(R)"        
#> [109] "NO3(R)"            "gL(R)"             "AL-DL4(R)"        
#> [112] "AL-VM7d(R)"        "SMP(R)"            "AL-VA6(R)"        
#> [115] "AL-VP1l(R)"        "FB"                "FBl6"             
#> [118] "IPS(R)"            "LX(L)"             "AL-DL2v(R)"       
#> [121] "b'L(R)"            "SNP(L)"            "gL(L)"            
#> [124] "AL-D(L)"           "FBl5"              "AL-DL2d(R)"       
#> [127] "PB(R6)"            "FBl4"              "mALT(L)"          
#> [130] "LX(R)"             "AL-DC4(L)"         "VES(R)"           
#> [133] "FBl7"              "AL-VL2p(R)"        "CA(R)"            
#> [136] "NO2(R)"            "AL-DA3(L)"         "AL-DM4(L)"        
#> [139] "b1(R)"             "SIP(L)"            "aL(R)"            
#> [142] "AL-DM1(L)"         "g5(R)"             "AL-DM4(R)"        
#> [145] "WED(R)"            "AL-VA7l(R)"        "AL-VL2a(R)"       
#> [148] "PENP"              "AOTU(R)"           "a'2(R)"           
#> [151] "AL-DA2(L)"         "AVLP(R)"           "LAL(L)"           
#> [154] "AL-DP1m(L)"        "AME(R)"            "GNG"              
#> [157] "EBr3am"            "FLA(R)"            "LAL(-GA)(R)"      
#> [160] "AL-DA4m(L)"        "PB(R9)"            "INP"              
#> [163] "AL-VM7d(L)"        "CRE(-RUB)(L)"      "ICL(L)"           
#> [166] "a'L(L)"            "AL-VM2(R)"         "PB(R8)"           
#> [169] "LAL(R)"            "VES(L)"            "AL-DL5(R)"        
#> [172] "AL-VL1(R)"         "a2(R)"             "AL-VA7m(R)"       
#> [175] "AL-VC3l(R)"        "AL-VP4(R)"         "NO1(R)"           
#> [178] "AL-DM6(L)"         "aL(L)"             "MB(R)"            
#> [181] "AL-D(R)"           "a1(R)"             "mALT(R)"          
#> [184] "AL-DL4(L)"         "SAD(-AMMC)"        "AL-VA4(R)"        
#> [187] "AL-VA1v(R)"        "AL-VM4(R)"         "AL-VP2(R)"        
#> [190] "POC"               "AMMC"              "hemibrain"        
#> [193] "SPS(L)"            "PLP(R)"            "PB(R2)"           
#> [196] "AL-DM2(R)"         "b'1(R)"            "GA(R)"            
#> [199] "EPA(L)"            "GOR(R)"            "AL-DC2(R)"        
#> [202] "AL-DC3(R)"         "FB-column3"        "AL-VC2(R)"        
#> [205] "g1(R)"             "EBr3pw"            "ATL(R)"           
#> [208] "VLNP(R)"           "MB(L)"             "AL-DC1(L)"        
#> [211] "PVLP(R)"           "AL-VA1d(R)"        "NO"               
#> [214] "PB(L5)"            "CRE(-ROB,-RUB)(R)" "SIP(R)"           
#> [217] "CRE(L)"            "IB112"             "AL-VC5(R)"        
#> [220] "LO(R)"             "AL-DP1m(R)"        "AL-VP5(R)"        
#> [223] "PRW"               "dACA(R)"           "a'L(R)"           
#> [226] "g4(R)"             "PB(L9)"            "AL-VP1d(R)"       
#> [229] "EBr5"              "PB(L8)"            "ATL(L)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "SCL(R)"            "dACA(R)"           "LX(R)"            
#>   [4] "EBr2r4"            "a3(R)"             "CX"               
#>   [7] "SPS(R)"            "AL(L)"             "VLNP(R)"          
#>  [10] "a1(R)"             "NO3(R)"            "LO(R)"            
#>  [13] "MB(+ACA)(R)"       "AB(L)"             "ATL(L)"           
#>  [16] "SAD"               "PRW"               "PB(R2)"           
#>  [19] "GNG"               "b'L(L)"            "PB(L5)"           
#>  [22] "g5(R)"             "AB(R)"             "PLP(R)"           
#>  [25] "a'1(R)"            "SLP(R)"            "RUB(R)"           
#>  [28] "PB(R6)"            "EBr1"              "PB(L8)"           
#>  [31] "PB(L9)"            "VES(R)"            "FBl5"             
#>  [34] "FBl8"              "PB"                "SNP(L)"           
#>  [37] "FBl2"              "b2(R)"             "VMNP"             
#>  [40] "g4(R)"             "ICL(L)"            "ICL(R)"           
#>  [43] "IPS(R)"            "MB(R)"             "a'3(R)"           
#>  [46] "SIP(L)"            "PB(R1)"            "FBl7"             
#>  [49] "mALT(L)"           "NO1(L)"            "PB(R9)"           
#>  [52] "NO(R)"             "MB(L)"             "FBl6"             
#>  [55] "LH(R)"             "VES(L)"            "b'1(R)"           
#>  [58] "LAL(L)"            "aL(R)"             "EB"               
#>  [61] "NO(L)"             "PB(L3)"            "LAL(-GA)(R)"      
#>  [64] "NO"                "FLA(R)"            "PB(R3)"           
#>  [67] "PB(L4)"            "a'L(L)"            "PENP"             
#>  [70] "EBr5"              "PB(R8)"            "PB(L7)"           
#>  [73] "AME(R)"            "LX(L)"             "OL(R)"            
#>  [76] "SMP(R)"            "g3(R)"             "GC"               
#>  [79] "SIP(R)"            "PVLP(R)"           "FBl4"             
#>  [82] "AVLP(R)"           "AMMC"              "b'2(R)"           
#>  [85] "lACA(R)"           "mALT(R)"           "NO1(R)"           
#>  [88] "a'2(R)"            "a'L(R)"            "gL(L)"            
#>  [91] "CRE(L)"            "PB(R5)"            "AOTU(R)"          
#>  [94] "b'L(R)"            "BU(R)"             "ATL(R)"           
#>  [97] "NO2(R)"            "RUB(L)"            "LAL(R)"           
#> [100] "AL(R)"             "EBr3am"            "g1(R)"            
#> [103] "PB(L6)"            "aL(L)"             "INP"              
#> [106] "SMP(L)"            "GOR(L)"            "NO3(L)"           
#> [109] "PB(L1)"            "CRE(-ROB,-RUB)(R)" "CA(L)"            
#> [112] "GOR(R)"            "hemibrain"         "CRE(R)"           
#> [115] "BU(L)"             "EBr3d"             "AOT(R)"           
#> [118] "PED(R)"            "FBl3"              "EPA(L)"           
#> [121] "PB(R4)"            "POC"               "NO2(L)"           
#> [124] "a2(R)"             "WED(R)"            "FB"               
#> [127] "CA(R)"             "CAN(R)"            "IB"               
#> [130] "LOP(R)"            "EBr6"              "SPS(L)"           
#> [133] "bL(L)"             "SNP(R)"            "ROB(R)"           
#> [136] "EBr3pw"            "FBl9"              "bL(R)"            
#> [139] "PB(L2)"            "vACA(R)"           "GF(R)"            
#> [142] "ME(R)"             "EPA(R)"            "gL(R)"            
#> [145] "GA(R)"             "FBl1"              "b1(R)"            
#> [148] "PB(R7)"            "g2(R)"             "SAD(-AMMC)"       
#> [151] "SCL(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DM3(R)"         "CRE(-ROB,-RUB)(R)" "AOTU(R)"          
#>   [4] "PB(R9)"            "AVLP(R)"           "AL-VA1d(R)"       
#>   [7] "AL-DP1m(L)"        "AL-VA7m(R)"        "RUB(R)"           
#>  [10] "PB(L2)"            "FB-column3"        "AL-VM7v(R)"       
#>  [13] "AB(L)"             "PB(R8)"            "AL-VC4(R)"        
#>  [16] "GA(R)"             "lACA(R)"           "g4(R)"            
#>  [19] "AL-DC3(R)"         "EB"                "MB(R)"            
#>  [22] "AL-DA2(R)"         "a3(R)"             "AL-VM2(R)"        
#>  [25] "mALT(R)"           "PB(L8)"            "NO3(L)"           
#>  [28] "FBl4"              "AL-DA2(L)"         "PB(L9)"           
#>  [31] "LX(L)"             "AL(R)"             "AL-DA3(R)"        
#>  [34] "AL-VA5(R)"         "PRW"               "aL(R)"            
#>  [37] "b'L(R)"            "AL-VM7v(L)"        "bL(L)"            
#>  [40] "GNG"               "b'L(L)"            "AMMC"             
#>  [43] "AL-DM5(L)"         "AL-VM7d(L)"        "b1(R)"            
#>  [46] "IPS(R)"            "PB(R6)"            "PB(L3)"           
#>  [49] "NO3(R)"            "AL-VA7l(R)"        "AL-D(L)"          
#>  [52] "LAL(R)"            "PB(L5)"            "LAL(-GA)(R)"      
#>  [55] "dACA(R)"           "BU(R)"             "PB(R4)"           
#>  [58] "AL-VP3(R)"         "PB(L6)"            "NO(R)"            
#>  [61] "FBl7"              "LH(R)"             "SPS(R)"           
#>  [64] "PB"                "LO(R)"             "AL-VM5d(R)"       
#>  [67] "RUB(L)"            "a'L(L)"            "AL-DL2v(R)"       
#>  [70] "AL-VL2a(R)"        "EBr5"              "AL-DM3(L)"        
#>  [73] "SIP(R)"            "PB(R3)"            "AL-V(R)"          
#>  [76] "CX"                "NO1(R)"            "AL-VL1(R)"        
#>  [79] "AL-VP1d(R)"        "CA(R)"             "AL-VP2(R)"        
#>  [82] "NO1(L)"            "VMNP"              "ROB(R)"           
#>  [85] "MB(+ACA)(R)"       "hemibrain"         "CRE(R)"           
#>  [88] "a'L(R)"            "AOT(R)"            "EBr3d"            
#>  [91] "NO"                "ATL(R)"            "PB(L4)"           
#>  [94] "PB(L1)"            "IB"                "LOP(R)"           
#>  [97] "FLA(R)"            "FBl2"              "vACA(R)"          
#> [100] "SCL(L)"            "SIP(L)"            "WED(R)"           
#> [103] "AL-DM2(R)"         "AL-VM5v(R)"        "a2(R)"            
#> [106] "EBr1"              "AL-VC2(R)"         "MB(L)"            
#> [109] "AL-DL2d(R)"        "AL-DP1l(R)"        "FBl3"             
#> [112] "EPA(L)"            "NO(L)"             "AL-DC2(R)"        
#> [115] "SAD"               "g2(R)"             "FBl8"             
#> [118] "CA(L)"             "ME(R)"             "AL-VA1v(R)"       
#> [121] "AL-VC3m(R)"        "AL-VP1m(R)"        "AL-DA3(L)"        
#> [124] "PVLP(R)"           "NO2(R)"            "FB"               
#> [127] "SCL(R)"            "a'2(R)"            "AL-VA6(R)"        
#> [130] "g3(R)"             "gL(L)"             "SNP(L)"           
#> [133] "AL-DA4m(L)"        "CRE(L)"            "SMP(R)"           
#> [136] "AL-VM7d(R)"        "NO2(L)"            "GF(R)"            
#> [139] "ICL(L)"            "FBl1"              "AL-DM1(R)"        
#> [142] "bL(R)"             "AL-D(R)"           "AL-DM6(L)"        
#> [145] "INP"               "AL-VA2(R)"         "ATL(L)"           
#> [148] "EBr6"              "AL-DC1(L)"         "AB(R)"            
#> [151] "AL-VP4(R)"         "SAD(-AMMC)"        "AL-DC1(R)"        
#> [154] "AL-DL1(R)"         "VLNP(R)"           "AL-DA4m(R)"       
#> [157] "LAL(L)"            "g1(R)"             "VES(L)"           
#> [160] "AL-DA1(R)"         "a'1(R)"            "FBl6"             
#> [163] "AL-DL3(R)"         "AL-DL4(R)"         "AL-DP1m(R)"       
#> [166] "EBr3am"            "SNP(R)"            "AL-DC4(R)"        
#> [169] "AL-DM6(R)"         "AL-VL2p(R)"        "PB(R7)"           
#> [172] "AL(L)"             "b'1(R)"            "PENP"             
#> [175] "AL-VA3(R)"         "CAN(R)"            "a'3(R)"           
#> [178] "EBr2r4"            "AL-DC2(L)"         "AL-DL4(L)"        
#> [181] "GOR(R)"            "PED(R)"            "b2(R)"            
#> [184] "FBl9"              "mALT(L)"           "OL(R)"            
#> [187] "AL-VP1l(R)"        "a1(R)"             "AL-VC5(R)"        
#> [190] "AL-VM3(R)"         "PB(R5)"            "GOR(L)"           
#> [193] "GC"                "CRE(-RUB)(L)"      "AME(R)"           
#> [196] "gL(R)"             "AL-DA4l(R)"        "PLP(R)"           
#> [199] "g5(R)"             "FBl5"              "AL-VA6(L)"        
#> [202] "VES(R)"            "AL-DM4(L)"         "AL-DL5(L)"        
#> [205] "PB(L7)"            "AL-DL5(R)"         "POC"              
#> [208] "b'2(R)"            "PB(R2)"            "EPA(R)"           
#> [211] "AL-DC4(L)"         "ICL(R)"            "AL-VC3l(R)"       
#> [214] "AL-VC1(R)"         "AL-VM1(R)"         "AL-VP5(R)"        
#> [217] "SPS(L)"            "LX(R)"             "EBr3pw"           
#> [220] "AL-DM4(R)"         "BU(L)"             "AL-DM1(L)"        
#> [223] "AL-VA4(R)"         "aL(L)"             "SMP(L)"           
#> [226] "PB(R1)"            "AL-VM4(R)"         "AL-DM2(L)"        
#> [229] "SLP(R)"            "AL-DM5(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "EBr1"              "AL-DM5(R)"         "NO3(R)"           
#>   [4] "AL-VA3(R)"         "LOP(R)"            "AL-DL5(L)"        
#>   [7] "AL-DM5(L)"         "CAN(R)"            "dACA(R)"          
#>  [10] "AL(L)"             "AL-DP1m(L)"        "PB"               
#>  [13] "CA(R)"             "AL-DL4(R)"         "AL-VA6(L)"        
#>  [16] "NO1(L)"            "BU(L)"             "PB(L8)"           
#>  [19] "AL-VM7v(R)"        "OL(R)"             "AL-VL1(R)"        
#>  [22] "RUB(R)"            "EB"                "INP"              
#>  [25] "AL-DL3(R)"         "AL-DM4(L)"         "AL-VC1(R)"        
#>  [28] "PB(L5)"            "g1(R)"             "AL-DM3(L)"        
#>  [31] "AL-VA1v(R)"        "AL-DL1(R)"         "g2(R)"            
#>  [34] "NO1(R)"            "AOT(R)"            "GOR(L)"           
#>  [37] "mALT(R)"           "PB(L6)"            "AL-VA1d(R)"       
#>  [40] "PB(L3)"            "SPS(R)"            "PED(R)"           
#>  [43] "AL-DC4(L)"         "SIP(R)"            "AL-VL2p(R)"       
#>  [46] "AL-VM5d(R)"        "AL-DA4m(L)"        "bL(R)"            
#>  [49] "AL-VC3m(R)"        "LX(L)"             "VES(L)"           
#>  [52] "AL-D(R)"           "AL-VA7m(R)"        "GF(R)"            
#>  [55] "SAD"               "ME(R)"             "bL(L)"            
#>  [58] "MB(+ACA)(R)"       "FBl3"              "GOR(R)"           
#>  [61] "aL(L)"             "FB"                "FLA(R)"           
#>  [64] "AL-DC2(L)"         "AL-DM6(L)"         "a'L(L)"           
#>  [67] "AL-DA3(R)"         "PENP"              "a'2(R)"           
#>  [70] "AL-DC1(L)"         "CX"                "FBl6"             
#>  [73] "FBl7"              "AL-VM3(R)"         "AL-VA5(R)"        
#>  [76] "PB(R5)"            "a1(R)"             "a'1(R)"           
#>  [79] "AL-VA6(R)"         "AL-VC4(R)"         "POC"              
#>  [82] "AB(L)"             "SIP(L)"            "LX(R)"            
#>  [85] "g4(R)"             "a'3(R)"            "EBr6"             
#>  [88] "AL-VA4(R)"         "CRE(-ROB,-RUB)(R)" "AL-DM1(R)"        
#>  [91] "AL-VP1m(R)"        "AL-D(L)"           "NO3(L)"           
#>  [94] "EBr3pw"            "NO2(L)"            "MB(L)"            
#>  [97] "SPS(L)"            "AL-DC2(R)"         "b2(R)"            
#> [100] "AL-DM1(L)"         "LAL(L)"            "EBr2r4"           
#> [103] "AL-VL2a(R)"        "SCL(L)"            "AMMC"             
#> [106] "lACA(R)"           "PB(L9)"            "AL-DA2(R)"        
#> [109] "AL(R)"             "AL-DA1(R)"         "SNP(L)"           
#> [112] "PB(R2)"            "PB(R4)"            "AL-VM4(R)"        
#> [115] "SMP(R)"            "mALT(L)"           "NO2(R)"           
#> [118] "GNG"               "AL-DP1m(R)"        "LH(R)"            
#> [121] "AME(R)"            "AL-DM3(R)"         "PB(R8)"           
#> [124] "AL-VM7d(L)"        "PB(R1)"            "hemibrain"        
#> [127] "AL-VC3l(R)"        "MB(R)"             "AL-DM6(R)"        
#> [130] "AL-VP1l(R)"        "PRW"               "aL(R)"            
#> [133] "FBl9"              "LO(R)"             "AL-DL5(R)"        
#> [136] "SAD(-AMMC)"        "PB(L4)"            "SLP(R)"           
#> [139] "AL-VP5(R)"         "AL-VP4(R)"         "SMP(L)"           
#> [142] "a2(R)"             "EBr3d"             "VLNP(R)"          
#> [145] "gL(L)"             "AL-DA2(L)"         "AB(R)"            
#> [148] "GA(R)"             "NO"                "AL-DM4(R)"        
#> [151] "AL-VM7v(L)"        "ROB(R)"            "LAL(R)"           
#> [154] "EPA(L)"            "NO(L)"             "CA(L)"            
#> [157] "AOTU(R)"           "CRE(R)"            "CRE(L)"           
#> [160] "AL-DM2(L)"         "FBl2"              "AL-VA2(R)"        
#> [163] "IB"                "NO(R)"             "VES(R)"           
#> [166] "FBl1"              "AL-DL4(L)"         "CRE(-RUB)(L)"     
#> [169] "AL-VP2(R)"         "PB(R7)"            "ATL(R)"           
#> [172] "AL-DA4m(R)"        "AL-VM2(R)"         "vACA(R)"          
#> [175] "b'L(R)"            "PVLP(R)"           "RUB(L)"           
#> [178] "FBl4"              "AVLP(R)"           "AL-VM1(R)"        
#> [181] "AL-DM2(R)"         "AL-VM7d(R)"        "a'L(R)"           
#> [184] "gL(R)"             "ICL(L)"            "b1(R)"            
#> [187] "AL-V(R)"           "AL-VP1d(R)"        "EPA(R)"           
#> [190] "AL-DA3(L)"         "FBl8"              "AL-DL2d(R)"       
#> [193] "PB(R9)"            "AL-DP1l(R)"        "AL-VC5(R)"        
#> [196] "AL-VP3(R)"         "EBr3am"            "AL-VA7l(R)"       
#> [199] "ATL(L)"            "PB(R6)"            "b'2(R)"           
#> [202] "EBr5"              "SNP(R)"            "ICL(R)"           
#> [205] "IPS(R)"            "WED(R)"            "PB(L1)"           
#> [208] "PB(L7)"            "PB(R3)"            "a3(R)"            
#> [211] "LAL(-GA)(R)"       "SCL(R)"            "g5(R)"            
#> [214] "FBl5"              "AL-DL2v(R)"        "AL-VC2(R)"        
#> [217] "FB-column3"        "AL-DC4(R)"         "AL-VM5v(R)"       
#> [220] "g3(R)"             "BU(R)"             "AL-DA4l(R)"       
#> [223] "AL-DC1(R)"         "AL-DC3(R)"         "b'1(R)"           
#> [226] "GC"                "PLP(R)"            "b'L(L)"           
#> [229] "VMNP"              "PB(L2)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method   path                                   
#> [1,] "GET"    "/api/npexplorer/autapses"             
#> [2,] "GET"    "/profile"                             
#> [3,] "GET"    "/api"                                 
#> [4,] "DELETE" "/api/*"                               
#> [5,] "GET"    "/api/v:ver/npexplorer/neuronmeta"     
#> [6,] "GET"    "/api/v:ver/npexplorer/roiconnectivity"
#>      name                                                                                    
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getAutapses-fm"
#> [2,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                
#> [3,] "github.com/labstack/echo.glob..func1"                                                  
#> [4,] "github.com/labstack/echo.glob..func1"                                                  
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }