Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "AOT(R)"            "LAL(L)"            "SPS(R)"           
#>   [4] "NO3(R)"            "PB(R8)"            "LX(L)"            
#>   [7] "b'2(R)"            "NO(R)"             "PB(L2)"           
#>  [10] "CA(L)"             "a'L(R)"            "AB(R)"            
#>  [13] "PB(R2)"            "VES(L)"            "PB(R5)"           
#>  [16] "PB(L1)"            "gL(L)"             "PB(R9)"           
#>  [19] "g5(R)"             "FBl9"              "SCL(L)"           
#>  [22] "CX"                "PB(R6)"            "b2(R)"            
#>  [25] "LX(R)"             "SIP(R)"            "PB(R3)"           
#>  [28] "a2(R)"             "WED(R)"            "GOR(L)"           
#>  [31] "NO"                "LH(R)"             "MB(+ACA)(R)"      
#>  [34] "CRE(-ROB,-RUB)(R)" "AMMC"              "g2(R)"            
#>  [37] "SNP(R)"            "EPA(R)"            "CRE(L)"           
#>  [40] "EBr3pw"            "aL(R)"             "FB"               
#>  [43] "PB(L6)"            "FBl7"              "FBl5"             
#>  [46] "ATL(R)"            "AVLP(R)"           "PB(L3)"           
#>  [49] "LAL(-GA)(R)"       "EBr6"              "ME(R)"            
#>  [52] "SNP(L)"            "OL(R)"             "EBr3am"           
#>  [55] "NO2(L)"            "PB(R1)"            "PB(L7)"           
#>  [58] "PVLP(R)"           "FBl2"              "EPA(L)"           
#>  [61] "LAL(R)"            "hemibrain"         "ICL(L)"           
#>  [64] "IPS(R)"            "IB"                "a'3(R)"           
#>  [67] "SAD"               "NO2(R)"            "RUB(L)"           
#>  [70] "LOP(R)"            "MB(R)"             "GF(R)"            
#>  [73] "FBl6"              "a3(R)"             "PB(L4)"           
#>  [76] "a'L(L)"            "GA(R)"             "EBr5"             
#>  [79] "g4(R)"             "CA(R)"             "g3(R)"            
#>  [82] "PED(R)"            "GOR(R)"            "b'1(R)"           
#>  [85] "PB(L5)"            "a1(R)"             "PENP"             
#>  [88] "AOTU(R)"           "a'2(R)"            "EB"               
#>  [91] "AL(R)"             "VLNP(R)"           "MB(L)"            
#>  [94] "a'1(R)"            "LO(R)"             "b1(R)"            
#>  [97] "CAN(R)"            "GC"                "SLP(R)"           
#> [100] "SMP(L)"            "FBl1"              "FBl4"             
#> [103] "bL(R)"             "POC"               "lACA(R)"          
#> [106] "b'L(R)"            "BU(R)"             "FBl3"             
#> [109] "mALT(L)"           "PB(R7)"            "VES(R)"           
#> [112] "aL(L)"             "PRW"               "PB(L8)"           
#> [115] "NO1(R)"            "ROB(R)"            "FBl8"             
#> [118] "NO(L)"             "AB(L)"             "BU(L)"            
#> [121] "FLA(R)"            "AME(R)"            "INP"              
#> [124] "GNG"               "PB"                "gL(R)"            
#> [127] "SIP(L)"            "SCL(R)"            "SPS(L)"           
#> [130] "SAD(-AMMC)"        "NO3(L)"            "mALT(R)"          
#> [133] "NO1(L)"            "CRE(R)"            "EBr1"             
#> [136] "PB(L9)"            "b'L(L)"            "SMP(R)"           
#> [139] "RUB(R)"            "PLP(R)"            "VMNP"             
#> [142] "EBr3d"             "AL(L)"             "ICL(R)"           
#> [145] "vACA(R)"           "ATL(L)"            "g1(R)"            
#> [148] "dACA(R)"           "EBr2r4"            "bL(L)"            
#> [151] "PB(R4)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                                  
#> [1,] "CONNECT" "/api/*"                              
#> [2,] "GET"     "/api/v:ver/dbmeta/version"           
#> [3,] "GET"     "/api/v:ver/dbmeta/instances"         
#> [4,] "GET"     "/api/v:ver/npexplorer/neuronmetavals"
#> [5,] "GET"     "/api/v:ver/npexplorer/rankedtable"   
#> [6,] "POST"    "/api/v:ver/npexplorer/rankedtable"   
#>      name                                                                
#> [1,] "github.com/labstack/echo.glob..func1"                              
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }