Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0`
#> $`hemibrain:v1.0`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0`$ROIs
#>   [1] "OL(R)"             "LAL(L)"            "b'L(R)"           
#>   [4] "FBl8"              "g2(R)"             "NO1(R)"           
#>   [7] "EPA(R)"            "CRE(L)"            "VES(L)"           
#>  [10] "VMNP"              "a'L(R)"            "NO"               
#>  [13] "SMP(R)"            "FBl7"              "CAN(R)"           
#>  [16] "EPA(L)"            "MB(R)"             "AOT(R)"           
#>  [19] "hemibrain"         "bL(L)"             "INP"              
#>  [22] "PB(L5)"            "SCL(R)"            "gL(R)"            
#>  [25] "vACA(R)"           "LAL(-GA)(R)"       "LO(R)"            
#>  [28] "PED(R)"            "MB(L)"             "PVLP(R)"          
#>  [31] "GA(R)"             "PB(R7)"            "EBr3d"            
#>  [34] "AVLP(R)"           "RUB(R)"            "b'1(R)"           
#>  [37] "AMMC"              "SAD"               "PB(R8)"           
#>  [40] "NO3(R)"            "ME(R)"             "PB(L2)"           
#>  [43] "PB(R6)"            "b'2(R)"            "SAD(-AMMC)"       
#>  [46] "LAL(R)"            "a1(R)"             "PB(L9)"           
#>  [49] "NO3(L)"            "NO2(L)"            "CRE(R)"           
#>  [52] "PB(R4)"            "PB"                "g4(R)"            
#>  [55] "EBr3pw"            "LH(R)"             "b1(R)"            
#>  [58] "AB(R)"             "SMP(L)"            "FLA(R)"           
#>  [61] "AB(L)"             "mALT(L)"           "FBl9"             
#>  [64] "VLNP(R)"           "PB(L1)"            "VES(R)"           
#>  [67] "AME(R)"            "NO2(R)"            "a'3(R)"           
#>  [70] "GF(R)"             "SPS(L)"            "PB(R5)"           
#>  [73] "PB(L6)"            "FBl5"              "SCL(L)"           
#>  [76] "EBr2r4"            "a'1(R)"            "GC"               
#>  [79] "SIP(R)"            "PB(R1)"            "FBl6"             
#>  [82] "PB(L4)"            "AL(R)"             "PRW"              
#>  [85] "mALT(R)"           "CX"                "POC"              
#>  [88] "ATL(L)"            "EBr1"              "BU(R)"            
#>  [91] "FBl1"              "RUB(L)"            "AOTU(R)"          
#>  [94] "g1(R)"             "PB(R2)"            "LOP(R)"           
#>  [97] "CRE(-ROB,-RUB)(R)" "EB"                "PB(R9)"           
#> [100] "SPS(R)"            "PLP(R)"            "LX(L)"            
#> [103] "MB(+ACA)(R)"       "ATL(R)"            "IPS(R)"           
#> [106] "IB"                "PB(L8)"            "SNP(L)"           
#> [109] "FBl2"              "FBl3"              "bL(R)"            
#> [112] "SIP(L)"            "WED(R)"            "b2(R)"            
#> [115] "FB"                "BU(L)"             "SNP(R)"           
#> [118] "PB(R3)"            "GOR(L)"            "aL(R)"            
#> [121] "g5(R)"             "aL(L)"             "PB(L3)"           
#> [124] "ROB(R)"            "lACA(R)"           "PB(L7)"           
#> [127] "g3(R)"             "a2(R)"             "a'2(R)"           
#> [130] "gL(L)"             "AL(L)"             "ICL(L)"           
#> [133] "FBl4"              "ICL(R)"            "NO1(L)"           
#> [136] "EBr5"              "GOR(R)"            "SLP(R)"           
#> [139] "a3(R)"             "a'L(L)"            "NO(R)"            
#> [142] "GNG"               "EBr3am"            "CA(L)"            
#> [145] "PENP"              "LX(R)"             "EBr6"             
#> [148] "NO(L)"             "b'L(L)"            "dACA(R)"          
#> [151] "CA(R)"            
#> 
#> $`hemibrain:v1.0`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                     
#> [1,] "POST" "/api/v:ver/npexplorer/simpleconnections"
#> [2,] "GET"  "/api/v:ver/npexplorer/rankedtable"      
#> [3,] "POST" "/api/npexplorer/commonconnectivity"     
#> [4,] "POST" "/api/npexplorer/celltype/:dataset/:type"
#> [5,] "GET"  "/api/v:ver/cached/roiconnectivity"      
#> [6,] "GET"  "/api/v:ver/dbmeta/version"              
#>      name                                                                                              
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCommonConnectivity-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCellType-fm"          
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                              

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }