Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "lACA(R)"           "NO2(R)"            "PB(R4)"           
#>   [4] "AL-VM4(R)"         "CA(L)"             "AL(L)"            
#>   [7] "PB(R5)"            "a'3(R)"            "LAL(L)"           
#>  [10] "CA(R)"             "ATL(R)"            "LX(L)"            
#>  [13] "GF(R)"             "AL-DC4(L)"         "AL-DA4m(R)"       
#>  [16] "AL-VA2(R)"         "AL-VP1m(R)"        "PB(L8)"           
#>  [19] "GOR(R)"            "GNG"               "AL-DC4(R)"        
#>  [22] "LOP(R)"            "AOTU(R)"           "SNP(L)"           
#>  [25] "FBl1"              "FBl5"              "AL-VL1(R)"        
#>  [28] "g2(R)"             "EBr6"              "gL(R)"            
#>  [31] "vACA(R)"           "mALT(L)"           "a1(R)"            
#>  [34] "PB(L6)"            "AL-VP2(R)"         "SCL(R)"           
#>  [37] "EB"                "AL-DM5(L)"         "AL-VA6(L)"        
#>  [40] "AL(R)"             "AL-DA4l(R)"        "AL-DP1l(R)"       
#>  [43] "VLNP(R)"           "CX"                "hemibrain"        
#>  [46] "AL-DA2(L)"         "PB(R9)"            "aL(R)"            
#>  [49] "EPA(L)"            "AL-VA5(R)"         "IB"               
#>  [52] "CAN(R)"            "AL-VP3(R)"         "ROB(R)"           
#>  [55] "INP"               "RUB(L)"            "AL-VA4(R)"        
#>  [58] "AL-VM7v(R)"        "PB(L4)"            "PB(R8)"           
#>  [61] "LX(R)"             "EBr3d"             "FBl8"             
#>  [64] "SPS(R)"            "AL-DP1m(L)"        "b1(R)"            
#>  [67] "AL-VA6(R)"         "AL-DM2(R)"         "AL-DM5(R)"        
#>  [70] "LO(R)"             "AL-VC3m(R)"        "SAD"              
#>  [73] "b2(R)"             "FBl3"              "b'2(R)"           
#>  [76] "AL-DM4(L)"         "PB(L7)"            "AL-DM6(R)"        
#>  [79] "b'1(R)"            "FBl9"              "AL-DL5(R)"        
#>  [82] "AL-VA7m(R)"        "a'1(R)"            "AL-VP4(R)"        
#>  [85] "SNP(R)"            "OL(R)"             "AL-DP1m(R)"       
#>  [88] "AB(L)"             "EBr1"              "MB(R)"            
#>  [91] "AL-VM5d(R)"        "SIP(L)"            "AL-DL5(L)"        
#>  [94] "PB(R1)"            "EBr3pw"            "g3(R)"            
#>  [97] "a2(R)"             "PB(R7)"            "EBr5"             
#> [100] "FB-column3"        "AL-DC3(R)"         "AL-DL1(R)"        
#> [103] "SPS(L)"            "AL-DM1(L)"         "AL-DL3(R)"        
#> [106] "NO3(R)"            "NO1(R)"            "AL-DC1(L)"        
#> [109] "AL-DM6(L)"         "bL(L)"             "AL-VC3l(R)"       
#> [112] "AL-VL2a(R)"        "AL-VM5v(R)"        "AL-VM7d(R)"       
#> [115] "GC"                "g1(R)"             "VMNP"             
#> [118] "PB(R3)"            "PB(L2)"            "AL-DA1(R)"        
#> [121] "AL-DM4(R)"         "PENP"              "mALT(R)"          
#> [124] "BU(L)"             "gL(L)"             "FBl7"             
#> [127] "b'L(R)"            "FBl2"              "RUB(R)"           
#> [130] "PB(L5)"            "CRE(-ROB,-RUB)(R)" "ME(R)"            
#> [133] "AL-DL4(L)"         "AL-DM2(L)"         "PB(R2)"           
#> [136] "AL-DC2(L)"         "AL-D(L)"           "AL-DM3(R)"        
#> [139] "AL-VC2(R)"         "AL-VP5(R)"         "POC"              
#> [142] "PB(L3)"            "a'2(R)"            "AL-DA3(L)"        
#> [145] "SAD(-AMMC)"        "MB(L)"             "AB(R)"            
#> [148] "SMP(L)"            "AL-DL2d(R)"        "AOT(R)"           
#> [151] "FB"                "FLA(R)"            "AL-VC4(R)"        
#> [154] "EPA(R)"            "AVLP(R)"           "AL-DC1(R)"        
#> [157] "AL-VP1l(R)"        "LAL(-GA)(R)"       "WED(R)"           
#> [160] "PB"                "EBr2r4"            "BU(R)"            
#> [163] "AL-VM7v(L)"        "AME(R)"            "VES(R)"           
#> [166] "LAL(R)"            "AL-DA3(R)"         "AL-DL2v(R)"       
#> [169] "PB(L9)"            "ICL(L)"            "SMP(R)"           
#> [172] "g4(R)"             "NO1(L)"            "b'L(L)"           
#> [175] "PB(R6)"            "GA(R)"             "AL-VM7d(L)"       
#> [178] "CRE(L)"            "AL-DC2(R)"         "AL-VL2p(R)"       
#> [181] "PB(L1)"            "FBl6"              "LH(R)"            
#> [184] "SLP(R)"            "AL-D(R)"           "AL-VM3(R)"        
#> [187] "AL-DA4m(L)"        "ICL(R)"            "AL-DL4(R)"        
#> [190] "AL-V(R)"           "AMMC"              "dACA(R)"          
#> [193] "PLP(R)"            "NO"                "a3(R)"            
#> [196] "AL-VA7l(R)"        "AL-DM3(L)"         "FBl4"             
#> [199] "IPS(R)"            "CRE(R)"            "NO3(L)"           
#> [202] "MB(+ACA)(R)"       "AL-VM2(R)"         "PRW"              
#> [205] "GOR(L)"            "CRE(-RUB)(L)"      "a'L(R)"           
#> [208] "SIP(R)"            "g5(R)"             "bL(R)"            
#> [211] "AL-DM1(R)"         "AL-VA1d(R)"        "AL-VC5(R)"        
#> [214] "EBr3am"            "PED(R)"            "NO(R)"            
#> [217] "AL-VC1(R)"         "NO2(L)"            "a'L(L)"           
#> [220] "VES(L)"            "PVLP(R)"           "aL(L)"            
#> [223] "AL-DA2(R)"         "NO(L)"             "AL-VP1d(R)"       
#> [226] "IB112"             "AL-VA3(R)"         "SCL(L)"           
#> [229] "ATL(L)"            "AL-VA1v(R)"        "AL-VM1(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PVLP(R)"           "CAN(R)"            "a'L(L)"           
#>   [4] "mALT(R)"           "gL(L)"             "VES(L)"           
#>   [7] "LAL(L)"            "FBl2"              "EBr2r4"           
#>  [10] "AL(L)"             "PB(L6)"            "b1(R)"            
#>  [13] "RUB(R)"            "g1(R)"             "EBr1"             
#>  [16] "LOP(R)"            "b'2(R)"            "RUB(L)"           
#>  [19] "NO1(L)"            "PB(L3)"            "aL(R)"            
#>  [22] "FBl8"              "EB"                "LO(R)"            
#>  [25] "PB(R4)"            "CA(L)"             "PB(L9)"           
#>  [28] "EPA(L)"            "SLP(R)"            "MB(R)"            
#>  [31] "a'3(R)"            "PB(R6)"            "NO3(R)"           
#>  [34] "INP"               "PB(L4)"            "AB(L)"            
#>  [37] "AOTU(R)"           "aL(L)"             "PB(R1)"           
#>  [40] "a'1(R)"            "ATL(L)"            "b2(R)"            
#>  [43] "PB"                "EBr3pw"            "VES(R)"           
#>  [46] "PB(R3)"            "LAL(-GA)(R)"       "PB(R8)"           
#>  [49] "FBl7"              "LX(L)"             "LAL(R)"           
#>  [52] "AOT(R)"            "PB(L7)"            "FBl6"             
#>  [55] "PB(R2)"            "ATL(R)"            "vACA(R)"          
#>  [58] "PLP(R)"            "NO(R)"             "NO"               
#>  [61] "bL(L)"             "NO3(L)"            "FBl4"             
#>  [64] "FBl5"              "PENP"              "BU(L)"            
#>  [67] "dACA(R)"           "SNP(L)"            "CX"               
#>  [70] "SIP(L)"            "SCL(R)"            "EBr6"             
#>  [73] "g4(R)"             "AB(R)"             "GOR(R)"           
#>  [76] "GF(R)"             "WED(R)"            "FB"               
#>  [79] "SCL(L)"            "PB(L5)"            "mALT(L)"          
#>  [82] "SPS(R)"            "GOR(L)"            "CRE(L)"           
#>  [85] "GNG"               "bL(R)"             "IB"               
#>  [88] "PB(L1)"            "PB(L8)"            "a1(R)"            
#>  [91] "lACA(R)"           "PB(R5)"            "CRE(R)"           
#>  [94] "MB(L)"             "SPS(L)"            "a'2(R)"           
#>  [97] "EPA(R)"            "ROB(R)"            "hemibrain"        
#> [100] "SAD"               "FBl3"              "VLNP(R)"          
#> [103] "EBr5"              "NO1(R)"            "AME(R)"           
#> [106] "NO2(R)"            "NO(L)"             "SMP(L)"           
#> [109] "g2(R)"             "b'L(R)"            "g5(R)"            
#> [112] "FBl1"              "AVLP(R)"           "EBr3am"           
#> [115] "PED(R)"            "GC"                "ICL(L)"           
#> [118] "SAD(-AMMC)"        "FBl9"              "CRE(-ROB,-RUB)(R)"
#> [121] "FLA(R)"            "BU(R)"             "ME(R)"            
#> [124] "VMNP"              "LX(R)"             "SNP(R)"           
#> [127] "gL(R)"             "a3(R)"             "AL(R)"            
#> [130] "a2(R)"             "PB(R7)"            "LH(R)"            
#> [133] "ICL(R)"            "b'L(L)"            "NO2(L)"           
#> [136] "GA(R)"             "MB(+ACA)(R)"       "PB(R9)"           
#> [139] "OL(R)"             "PRW"               "SIP(R)"           
#> [142] "b'1(R)"            "PB(L2)"            "EBr3d"            
#> [145] "SMP(R)"            "CA(R)"             "IPS(R)"           
#> [148] "POC"               "AMMC"              "g3(R)"            
#> [151] "a'L(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "NO1(R)"            "ATL(R)"            "PB(L4)"           
#>   [4] "AL-DC4(R)"         "SAD"               "VES(R)"           
#>   [7] "AL(R)"             "VMNP"              "AL-DA4m(L)"       
#>  [10] "AL-DP1m(L)"        "AL-VM5d(R)"        "EBr3d"            
#>  [13] "a'L(R)"            "AL-DC3(R)"         "PB(R6)"           
#>  [16] "b2(R)"             "dACA(R)"           "FBl3"             
#>  [19] "CAN(R)"            "LAL(R)"            "AL-VA6(R)"        
#>  [22] "EBr5"              "FLA(R)"            "SNP(L)"           
#>  [25] "AL-VL2a(R)"        "lACA(R)"           "LAL(-GA)(R)"      
#>  [28] "GC"                "AL-VA1d(R)"        "AL-VP5(R)"        
#>  [31] "WED(R)"            "FBl8"              "AL-VP1l(R)"       
#>  [34] "AL-DL2d(R)"        "AL-VA3(R)"         "AL-VL1(R)"        
#>  [37] "CRE(-ROB,-RUB)(R)" "AL-DC4(L)"         "AL-DM1(L)"        
#>  [40] "aL(L)"             "FB"                "FB-column3"       
#>  [43] "AL-DC1(R)"         "SCL(L)"            "NO(R)"            
#>  [46] "AL-VC2(R)"         "MB(L)"             "AB(R)"            
#>  [49] "AL-DM6(R)"         "AL-VM7v(R)"        "NO1(L)"           
#>  [52] "CRE(-RUB)(L)"      "PB"                "LO(R)"            
#>  [55] "AVLP(R)"           "b'L(L)"            "AL-DM1(R)"        
#>  [58] "PB(L2)"            "a1(R)"             "PB(L3)"           
#>  [61] "AL-D(L)"           "AL-VA2(R)"         "AL-VC3m(R)"       
#>  [64] "AL-VM3(R)"         "AL-VP3(R)"         "PB(R1)"           
#>  [67] "ICL(L)"            "GNG"               "GOR(R)"           
#>  [70] "VES(L)"            "SMP(R)"            "NO2(L)"           
#>  [73] "PRW"               "PB(R7)"            "mALT(R)"          
#>  [76] "PB(L8)"            "EBr2r4"            "AL-DC2(L)"        
#>  [79] "CRE(L)"            "RUB(L)"            "AL-DA3(R)"        
#>  [82] "NO2(R)"            "AL-VM5v(R)"        "g4(R)"            
#>  [85] "SAD(-AMMC)"        "AL-VC3l(R)"        "AL-DM5(L)"        
#>  [88] "AL-VL2p(R)"        "AL(L)"             "aL(R)"            
#>  [91] "g5(R)"             "AL-DA4l(R)"        "GA(R)"            
#>  [94] "gL(L)"             "AL-DM6(L)"         "b'2(R)"           
#>  [97] "PLP(R)"            "bL(L)"             "LH(R)"            
#> [100] "PB(R3)"            "IPS(R)"            "SCL(R)"           
#> [103] "SNP(R)"            "AL-DM4(L)"         "FBl2"             
#> [106] "FBl4"              "b1(R)"             "AL-DL2v(R)"       
#> [109] "AL-VA7l(R)"        "SIP(L)"            "AL-DL5(L)"        
#> [112] "FBl5"              "AME(R)"            "LAL(L)"           
#> [115] "b'L(R)"            "AL-DC2(R)"         "AL-DL1(R)"        
#> [118] "AL-DP1l(R)"        "BU(L)"             "a'2(R)"           
#> [121] "EBr3pw"            "SLP(R)"            "AL-DM2(R)"        
#> [124] "AL-VM2(R)"         "vACA(R)"           "AL-VA1v(R)"       
#> [127] "VLNP(R)"           "PB(R8)"            "AL-DC1(L)"        
#> [130] "AL-VM7v(L)"        "AL-DM3(R)"         "ATL(L)"           
#> [133] "AOTU(R)"           "NO"                "PVLP(R)"          
#> [136] "AL-DL4(R)"         "LX(R)"             "AL-DA3(L)"        
#> [139] "AL-DA4m(R)"        "AL-VA5(R)"         "hemibrain"        
#> [142] "PB(L1)"            "ME(R)"             "PB(L6)"           
#> [145] "PB(L9)"            "PB(L7)"            "AL-VA7m(R)"       
#> [148] "a3(R)"             "MB(R)"             "CX"               
#> [151] "SMP(L)"            "EPA(R)"            "SIP(R)"           
#> [154] "PB(R4)"            "AL-DL4(L)"         "AL-DA1(R)"        
#> [157] "AL-DL3(R)"         "SPS(R)"            "AL-DA2(L)"        
#> [160] "a'1(R)"            "AL-DP1m(R)"        "AL-VP2(R)"        
#> [163] "CA(L)"             "NO3(L)"            "PB(R5)"           
#> [166] "AB(L)"             "RUB(R)"            "g1(R)"            
#> [169] "EB"                "MB(+ACA)(R)"       "AL-VM7d(R)"       
#> [172] "BU(R)"             "ROB(R)"            "a'L(L)"           
#> [175] "AL-DA2(R)"         "AL-D(R)"           "PB(R2)"           
#> [178] "g3(R)"             "g2(R)"             "LX(L)"            
#> [181] "INP"               "AL-DM5(R)"         "AMMC"             
#> [184] "a2(R)"             "EPA(L)"            "AL-VP4(R)"        
#> [187] "IB"                "LOP(R)"            "PED(R)"           
#> [190] "EBr1"              "FBl1"              "CA(R)"            
#> [193] "AL-VC4(R)"         "b'1(R)"            "CRE(R)"           
#> [196] "AL-DM4(R)"         "AL-VC1(R)"         "PENP"             
#> [199] "AL-DM3(L)"         "FBl6"              "AL-V(R)"          
#> [202] "GF(R)"             "AL-VM1(R)"         "POC"              
#> [205] "PB(R9)"            "FBl7"              "AL-VA4(R)"        
#> [208] "a'3(R)"            "mALT(L)"           "gL(R)"            
#> [211] "ICL(R)"            "NO3(R)"            "SPS(L)"           
#> [214] "AL-DM2(L)"         "AL-VM7d(L)"        "bL(R)"            
#> [217] "AL-VM4(R)"         "AL-VP1m(R)"        "PB(L5)"           
#> [220] "GOR(L)"            "EBr6"              "AL-VA6(L)"        
#> [223] "AL-DL5(R)"         "AL-VP1d(R)"        "EBr3am"           
#> [226] "OL(R)"             "AL-VC5(R)"         "AOT(R)"           
#> [229] "FBl9"              "NO(L)"            
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AL-VA7m(R)"        "FBl8"              "AL-DP1l(R)"       
#>   [4] "g4(R)"             "a'1(R)"            "FBl4"             
#>   [7] "AL-VM7v(R)"        "PB(L5)"            "SMP(L)"           
#>  [10] "VMNP"              "hemibrain"         "SIP(L)"           
#>  [13] "AL-DM4(L)"         "PB(R2)"            "AL-DA1(R)"        
#>  [16] "AL-VA4(R)"         "AL-VP1m(R)"        "gL(L)"            
#>  [19] "AL-DC4(L)"         "AL-VC3l(R)"        "CRE(R)"           
#>  [22] "CAN(R)"            "RUB(L)"            "AL-DC4(R)"        
#>  [25] "AL-DL2d(R)"        "AL-DL2v(R)"        "PB(R5)"           
#>  [28] "AL-DM1(L)"         "PB(L9)"            "GOR(L)"           
#>  [31] "WED(R)"            "AL-DL5(L)"         "AL-DM5(L)"        
#>  [34] "AL-VM7d(L)"        "b1(R)"             "PB(R6)"           
#>  [37] "GA(R)"             "AL-VL2a(R)"        "ATL(L)"           
#>  [40] "AL-VA6(R)"         "IB"                "PB(L2)"           
#>  [43] "b'L(R)"            "a'L(R)"            "OL(R)"            
#>  [46] "CA(L)"             "FB"                "ICL(R)"           
#>  [49] "EBr3d"             "AL-VC3m(R)"        "SPS(R)"           
#>  [52] "SPS(L)"            "AL-DA4m(R)"        "EBr5"             
#>  [55] "AL-DP1m(L)"        "AL-DM2(R)"         "AL-VA1v(R)"       
#>  [58] "a'3(R)"            "bL(L)"             "AL-DA4l(R)"       
#>  [61] "AL-VP3(R)"         "CRE(-ROB,-RUB)(R)" "AL-D(L)"          
#>  [64] "AL-DM6(R)"         "aL(R)"             "a'L(L)"           
#>  [67] "dACA(R)"           "PB(R9)"            "LX(R)"            
#>  [70] "NO3(L)"            "AL-VM3(R)"         "g3(R)"            
#>  [73] "b2(R)"             "EBr2r4"            "ICL(L)"           
#>  [76] "POC"               "SMP(R)"            "PRW"              
#>  [79] "PB"                "AL-DL4(R)"         "PB(L8)"           
#>  [82] "LAL(R)"            "mALT(R)"           "PB(L1)"           
#>  [85] "AL-DL4(L)"         "g5(R)"             "PENP"             
#>  [88] "LAL(-GA)(R)"       "AL-VL1(R)"         "AL-V(R)"          
#>  [91] "AL-DC1(L)"         "AL-DC1(R)"         "PED(R)"           
#>  [94] "AL-VM4(R)"         "PB(R7)"            "AL-DL5(R)"        
#>  [97] "NO2(L)"            "NO"                "IPS(R)"           
#> [100] "a3(R)"             "b'1(R)"            "FBl2"             
#> [103] "AL-DP1m(R)"        "AL-VM1(R)"         "b'2(R)"           
#> [106] "NO3(R)"            "AL-VC4(R)"         "AL-VL2p(R)"       
#> [109] "SCL(L)"            "SNP(R)"            "GOR(R)"           
#> [112] "AL-DA3(R)"         "AL-DA4m(L)"        "AB(R)"            
#> [115] "SAD(-AMMC)"        "CA(R)"             "PB(R8)"           
#> [118] "MB(R)"             "AL-DC2(R)"         "AL-DM5(R)"        
#> [121] "AL-VM5v(R)"        "NO1(L)"            "FLA(R)"           
#> [124] "FBl6"              "AL-DM3(R)"         "AL-VP1l(R)"       
#> [127] "AL-DC2(L)"         "gL(R)"             "INP"              
#> [130] "EB"                "AL-DM2(L)"         "lACA(R)"          
#> [133] "AL-DM6(L)"         "AL-VM7v(L)"        "GNG"              
#> [136] "LAL(L)"            "SAD"               "FBl3"             
#> [139] "AL-VC5(R)"         "MB(L)"             "AL-VA6(L)"        
#> [142] "FBl5"              "MB(+ACA)(R)"       "AL-VM7d(R)"       
#> [145] "g1(R)"             "EBr6"              "AL-DL3(R)"        
#> [148] "AL-VM2(R)"         "SNP(L)"            "VES(R)"           
#> [151] "FB-column3"        "EBr1"              "AL-DM3(L)"        
#> [154] "AL(R)"             "RUB(R)"            "AME(R)"           
#> [157] "SLP(R)"            "bL(R)"             "b'L(L)"           
#> [160] "AL-DA2(L)"         "FBl1"              "AOTU(R)"          
#> [163] "BU(L)"             "PB(L7)"            "AL-VM5d(R)"       
#> [166] "EBr3am"            "LX(L)"             "NO(R)"            
#> [169] "EBr3pw"            "CX"                "a'2(R)"           
#> [172] "AL-DC3(R)"         "AL-DM1(R)"         "AL-DM4(R)"        
#> [175] "ROB(R)"            "PB(R3)"            "AL-VP2(R)"        
#> [178] "AVLP(R)"           "AL-VP5(R)"         "EPA(L)"           
#> [181] "PB(R4)"            "GF(R)"             "SIP(R)"           
#> [184] "AB(L)"             "AL(L)"             "VES(L)"           
#> [187] "ME(R)"             "NO2(R)"            "LH(R)"            
#> [190] "AL-DA3(L)"         "CRE(-RUB)(L)"      "vACA(R)"          
#> [193] "AL-VA7l(R)"        "AOT(R)"            "EPA(R)"           
#> [196] "LO(R)"             "FBl7"              "AL-VA2(R)"        
#> [199] "LOP(R)"            "AMMC"              "SCL(R)"           
#> [202] "PLP(R)"            "AL-VP1d(R)"        "CRE(L)"           
#> [205] "aL(L)"             "AL-VA3(R)"         "AL-VA5(R)"        
#> [208] "BU(R)"             "PB(R1)"            "NO(L)"            
#> [211] "AL-VA1d(R)"        "AL-VC2(R)"         "VLNP(R)"          
#> [214] "g2(R)"             "AL-D(R)"           "AL-VC1(R)"        
#> [217] "mALT(L)"           "PB(L3)"            "PB(L4)"           
#> [220] "AL-VP4(R)"         "a1(R)"             "PVLP(R)"          
#> [223] "AL-DA2(R)"         "NO1(R)"            "PB(L6)"           
#> [226] "FBl9"              "ATL(R)"            "AL-DL1(R)"        
#> [229] "a2(R)"             "GC"               
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method    path                                  
#> [1,] "CONNECT" "/api"                                
#> [2,] "POST"    "/api/v:ver/raw/cypher/transaction"   
#> [3,] "POST"    "/api/raw/cypher/transaction/:id/kill"
#> [4,] "GET"     "/api/cached/dailytype"               
#> [5,] "GET"     "/oauth2callback"                     
#> [6,] "POST"    "/api/*"                              
#>      name                                                                              
#> [1,] "github.com/labstack/echo/v4.glob..func1"                                         
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"              
#> [3,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"          
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getDailyType-fm"
#> [5,] "github.com/janelia-flyem/echo-secure.oauthCallbackHandler"                       
#> [6,] "github.com/labstack/echo/v4.glob..func1"                                         
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }