Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "MB(+ACA)(R)"       "aL(L)"             "a'L(L)"           
#>   [4] "VLNP(R)"           "b2(R)"             "PB(R9)"           
#>   [7] "FB"                "SCL(R)"            "FBl6"             
#>  [10] "VES(L)"            "PED(R)"            "g1(R)"            
#>  [13] "bL(L)"             "dACA(R)"           "PB(R8)"           
#>  [16] "AL(L)"             "SIP(R)"            "LH(R)"            
#>  [19] "ICL(R)"            "g3(R)"             "gL(L)"            
#>  [22] "g2(R)"             "LX(R)"             "LX(L)"            
#>  [25] "NO(L)"             "NO"                "VES(R)"           
#>  [28] "b'L(L)"            "SCL(L)"            "b'L(R)"           
#>  [31] "EBr1"              "BU(L)"             "gL(R)"            
#>  [34] "b'2(R)"            "SIP(L)"            "a'1(R)"           
#>  [37] "PB(R3)"            "AB(R)"             "AL(R)"            
#>  [40] "CRE(-ROB,-RUB)(R)" "ATL(L)"            "SNP(R)"           
#>  [43] "ROB(R)"            "EBr2r4"            "PB(R2)"           
#>  [46] "IPS(R)"            "INP"               "LAL(R)"           
#>  [49] "GF(R)"             "CRE(R)"            "SNP(L)"           
#>  [52] "FBl8"              "POC"               "a1(R)"            
#>  [55] "PB(L3)"            "a2(R)"             "SLP(R)"           
#>  [58] "CX"                "LAL(L)"            "a'L(R)"           
#>  [61] "lACA(R)"           "aL(R)"             "PLP(R)"           
#>  [64] "FBl4"              "CA(R)"             "OL(R)"            
#>  [67] "EBr6"              "PB"                "SMP(R)"           
#>  [70] "LAL(-GA)(R)"       "SPS(L)"            "PRW"              
#>  [73] "PB(L8)"            "VMNP"              "NO(R)"            
#>  [76] "PB(L9)"            "RUB(L)"            "hemibrain"        
#>  [79] "SAD"               "FBl7"              "EB"               
#>  [82] "AVLP(R)"           "vACA(R)"           "mALT(L)"          
#>  [85] "mALT(R)"           "EBr3d"             "a'3(R)"           
#>  [88] "b1(R)"             "PB(L2)"            "EBr3am"           
#>  [91] "a'2(R)"            "g4(R)"             "AMMC"             
#>  [94] "CRE(L)"            "bL(R)"             "FBl1"             
#>  [97] "FBl3"              "GNG"               "MB(R)"            
#> [100] "IB"                "AOTU(R)"           "EBr3pw"           
#> [103] "SAD(-AMMC)"        "FBl9"              "PB(L5)"           
#> [106] "FLA(R)"            "ME(R)"             "a3(R)"            
#> [109] "PB(R5)"            "AB(L)"             "PB(R1)"           
#> [112] "GOR(R)"            "NO3(R)"            "MB(L)"            
#> [115] "PB(L1)"            "RUB(R)"            "SMP(L)"           
#> [118] "WED(R)"            "PB(L6)"            "FBl2"             
#> [121] "AME(R)"            "NO2(R)"            "CAN(R)"           
#> [124] "b'1(R)"            "PB(R6)"            "SPS(R)"           
#> [127] "LOP(R)"            "PENP"              "NO1(R)"           
#> [130] "PB(L4)"            "CA(L)"             "NO3(L)"           
#> [133] "EPA(L)"            "ICL(L)"            "PB(L7)"           
#> [136] "g5(R)"             "LO(R)"             "FBl5"             
#> [139] "GOR(L)"            "BU(R)"             "EPA(R)"           
#> [142] "GC"                "EBr5"              "AOT(R)"           
#> [145] "PVLP(R)"           "PB(R4)"            "NO1(L)"           
#> [148] "ATL(R)"            "NO2(L)"            "GA(R)"            
#> [151] "PB(R7)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-VP1m(R)"        "SCL(L)"            "SIP(R)"           
#>   [4] "AL-DP1m(L)"        "FB-column3"        "AL-DA2(R)"        
#>   [7] "AL-VA5(R)"         "EB"                "PB(L9)"           
#>  [10] "LAL(R)"            "NO2(L)"            "FB"               
#>  [13] "AL-VA6(L)"         "PB(R9)"            "IPS(R)"           
#>  [16] "AME(R)"            "AL-DP1m(R)"        "b'2(R)"           
#>  [19] "mALT(L)"           "NO3(R)"            "LH(R)"            
#>  [22] "SAD"               "AL-DL5(L)"         "AL-DA3(R)"        
#>  [25] "AL-VA1d(R)"        "AL-VL1(R)"         "CA(L)"            
#>  [28] "BU(R)"             "AOTU(R)"           "SPS(R)"           
#>  [31] "AL-DM1(R)"         "AL-VC4(R)"         "AL-VM7v(R)"       
#>  [34] "AL-VP3(R)"         "PB(L3)"            "PB(R1)"           
#>  [37] "CRE(R)"            "AB(R)"             "AL-VC3m(R)"       
#>  [40] "AL-V(R)"           "g1(R)"             "dACA(R)"          
#>  [43] "AL-VC5(R)"         "GC"                "FBl5"             
#>  [46] "PB(R2)"            "AL-DL2d(R)"        "AL-VP2(R)"        
#>  [49] "b2(R)"             "PLP(R)"            "AL-VA2(R)"        
#>  [52] "CRE(-ROB,-RUB)(R)" "GA(R)"             "RUB(L)"           
#>  [55] "MB(R)"             "AL-DA4m(L)"        "SMP(L)"           
#>  [58] "VES(L)"            "AL-VL2a(R)"        "lACA(R)"          
#>  [61] "AL-DA4m(R)"        "AL-DM2(L)"         "ATL(L)"           
#>  [64] "EBr2r4"            "a'L(R)"            "PB(R4)"           
#>  [67] "AL-D(R)"           "b'1(R)"            "EBr6"             
#>  [70] "gL(R)"             "FBl7"              "NO(L)"            
#>  [73] "PRW"               "FBl9"              "AL-DL3(R)"        
#>  [76] "SMP(R)"            "FBl8"              "PB"               
#>  [79] "AL-DL2v(R)"        "AL-DP1l(R)"        "EBr1"             
#>  [82] "AL-VL2p(R)"        "AL-DM5(L)"         "FBl2"             
#>  [85] "a'L(L)"            "PED(R)"            "PB(R3)"           
#>  [88] "CAN(R)"            "AL-VM5d(R)"        "g3(R)"            
#>  [91] "a'1(R)"            "EPA(L)"            "bL(R)"            
#>  [94] "AL-VA7m(R)"        "AL-DM1(L)"         "AL-VM7d(L)"       
#>  [97] "CRE(-RUB)(L)"      "LO(R)"             "LX(L)"            
#> [100] "OL(R)"             "AL-VP4(R)"         "VLNP(R)"          
#> [103] "EPA(R)"            "AL-DC4(L)"         "AL-DM2(R)"        
#> [106] "AL-VP1d(R)"        "a1(R)"             "AL(L)"            
#> [109] "aL(R)"             "PB(R5)"            "RUB(R)"           
#> [112] "a2(R)"             "AL-DC2(R)"         "aL(L)"            
#> [115] "SNP(L)"            "AL-DM6(R)"         "AL-VM2(R)"        
#> [118] "a'3(R)"            "BU(L)"             "SNP(R)"           
#> [121] "AL-DC2(L)"         "NO"                "ICL(L)"           
#> [124] "FBl1"              "CA(R)"             "EBr5"             
#> [127] "EBr3pw"            "SLP(R)"            "AL-VA7l(R)"       
#> [130] "AB(L)"             "GF(R)"             "NO2(R)"           
#> [133] "CX"                "AL-DA3(L)"         "AL-DM4(L)"        
#> [136] "VMNP"              "ICL(R)"            "AL-VA3(R)"        
#> [139] "NO3(L)"            "MB(L)"             "a'2(R)"           
#> [142] "AL(R)"             "AL-DM5(R)"         "AL-VA1v(R)"       
#> [145] "AL-VC1(R)"         "NO1(L)"            "hemibrain"        
#> [148] "b'L(R)"            "AL-VP5(R)"         "PENP"             
#> [151] "AL-DC4(R)"         "b'L(L)"            "AL-VM7v(L)"       
#> [154] "AL-DM3(R)"         "SPS(L)"            "FLA(R)"           
#> [157] "AL-DM3(L)"         "CRE(L)"            "g2(R)"            
#> [160] "AL-VC3l(R)"        "PB(L1)"            "gL(L)"            
#> [163] "VES(R)"            "AVLP(R)"           "PB(L4)"           
#> [166] "AL-VA4(R)"         "AL-VM4(R)"         "WED(R)"           
#> [169] "AL-VM7d(R)"        "AL-DL5(R)"         "g4(R)"            
#> [172] "FBl3"              "AL-DC3(R)"         "AL-DM4(R)"        
#> [175] "AL-VC2(R)"         "ATL(R)"            "PB(R8)"           
#> [178] "AL-DA2(L)"         "AL-DM6(L)"         "GNG"              
#> [181] "SCL(R)"            "b1(R)"             "POC"              
#> [184] "EBr3am"            "PB(L7)"            "PB(L2)"           
#> [187] "LAL(L)"            "AL-DA1(R)"         "AL-D(L)"          
#> [190] "AL-DC1(R)"         "bL(L)"             "AL-DA4l(R)"       
#> [193] "LAL(-GA)(R)"       "a3(R)"             "SAD(-AMMC)"       
#> [196] "PVLP(R)"           "AL-VM5v(R)"        "PB(R6)"           
#> [199] "AL-DC1(L)"         "PB(L6)"            "NO(R)"            
#> [202] "MB(+ACA)(R)"       "AL-DL1(R)"         "LOP(R)"           
#> [205] "NO1(R)"            "ME(R)"             "EBr3d"            
#> [208] "INP"               "PB(L5)"            "LX(R)"            
#> [211] "ROB(R)"            "g5(R)"             "FBl6"             
#> [214] "GOR(R)"            "AOT(R)"            "SIP(L)"           
#> [217] "vACA(R)"           "PB(L8)"            "GOR(L)"           
#> [220] "FBl4"              "AL-DL4(R)"         "PB(R7)"           
#> [223] "AL-VA6(R)"         "AL-VM1(R)"         "AL-DL4(L)"        
#> [226] "AL-VP1l(R)"        "mALT(R)"           "IB"               
#> [229] "AMMC"              "AL-VM3(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                           
#> [1,] "POST" "/api/raw/cypher/transaction"                  
#> [2,] "GET"  "/api/dbmeta/version"                          
#> [3,] "GET"  "/api/npexplorer/simpleconnections"            
#> [4,] "POST" "/api/v:ver/npexplorer/celltype/:dataset/:type"
#> [5,] "GET"  "/api/npexplorer/rankedtable"                  
#> [6,] "GET"  "/api/npexplorer/distribution"                 
#>      name                                                                                             
#> [1,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                         
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getVersion-fm"                  
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getSimpleConnections-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                             
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getRankedTable-fm"      
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"     

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }