Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "NO3(R)"            "FLA(R)"            "EBr5"             
#>   [4] "AL-DC4(L)"         "LX(L)"             "GA(R)"            
#>   [7] "AL-DP1m(L)"        "vACA(R)"           "CRE(-ROB,-RUB)(R)"
#>  [10] "LAL(L)"            "AOTU(R)"           "b2(R)"            
#>  [13] "AL-VP1d(R)"        "NO(R)"             "EBr2r4"           
#>  [16] "AL-DM6(R)"         "PB(R9)"            "FBl8"             
#>  [19] "AL-DM5(L)"         "AB(R)"             "AL-DC1(L)"        
#>  [22] "FBl7"              "SCL(L)"            "LAL(-GA)(R)"      
#>  [25] "MB(+ACA)(R)"       "AL-DM2(R)"         "AL-VA1v(R)"       
#>  [28] "AL-VM3(R)"         "AL-DC4(R)"         "AL-DM3(R)"        
#>  [31] "AL-VC3l(R)"        "AB(L)"             "PB(R8)"           
#>  [34] "gL(L)"             "FB-column3"        "FBl3"             
#>  [37] "CX"                "PED(R)"            "a'2(R)"           
#>  [40] "NO2(R)"            "INP"               "BU(R)"            
#>  [43] "NO"                "AL-DC2(R)"         "LAL(R)"           
#>  [46] "AL-VC5(R)"         "VLNP(R)"           "SIP(L)"           
#>  [49] "PB(R7)"            "VES(R)"            "a'L(L)"           
#>  [52] "AL-DM1(L)"         "b1(R)"             "FBl1"             
#>  [55] "PB(R3)"            "AL-DP1m(R)"        "LOP(R)"           
#>  [58] "FB"                "AL-VA6(L)"         "a'1(R)"           
#>  [61] "IPS(R)"            "PB(R6)"            "AL-VM2(R)"        
#>  [64] "CA(L)"             "aL(R)"             "GNG"              
#>  [67] "OL(R)"             "AL-VM5v(R)"        "PENP"             
#>  [70] "FBl6"              "ICL(R)"            "EPA(L)"           
#>  [73] "AL-DA2(R)"         "AL-DL2d(R)"        "PB(L3)"           
#>  [76] "MB(L)"             "AL-DL5(L)"         "AVLP(R)"          
#>  [79] "AL-DM4(R)"         "AL-DM5(R)"         "AL-D(R)"          
#>  [82] "AL-VL2p(R)"        "EPA(R)"            "AL-DM6(L)"        
#>  [85] "bL(R)"             "AL-DM1(R)"         "AL-VP2(R)"        
#>  [88] "RUB(R)"            "SPS(L)"            "CRE(L)"           
#>  [91] "AL-VA7m(R)"        "AL-VM1(R)"         "ATL(R)"           
#>  [94] "AL-VA1d(R)"        "AL-VM7v(R)"        "mALT(L)"          
#>  [97] "SNP(R)"            "AL-DA3(L)"         "AL-DA4m(L)"       
#> [100] "AL-DM2(L)"         "PB(R4)"            "AL-VM7d(R)"       
#> [103] "bL(L)"             "AL-VA6(R)"         "AL-VC1(R)"        
#> [106] "AL-VC3m(R)"        "gL(R)"             "PB(L9)"           
#> [109] "AL-D(L)"           "AME(R)"            "AL-VL1(R)"        
#> [112] "a1(R)"             "b'L(R)"            "AL-DC2(L)"        
#> [115] "a3(R)"             "AL-DL4(R)"         "b'1(R)"           
#> [118] "EBr6"              "BU(L)"             "g4(R)"            
#> [121] "PB(L4)"            "PB"                "IB112"            
#> [124] "AL-VM4(R)"         "SAD"               "PVLP(R)"          
#> [127] "PB(R5)"            "b'2(R)"            "g3(R)"            
#> [130] "SCL(R)"            "a'L(R)"            "AL-VC4(R)"        
#> [133] "EBr1"              "dACA(R)"           "EBr3d"            
#> [136] "CA(R)"             "NO1(L)"            "AL-DP1l(R)"       
#> [139] "AL-VA5(R)"         "CRE(R)"            "ME(R)"            
#> [142] "CAN(R)"            "AL-DL2v(R)"        "AL(L)"            
#> [145] "SLP(R)"            "AOT(R)"            "SIP(R)"           
#> [148] "a'3(R)"            "PRW"               "GC"               
#> [151] "VMNP"              "AL-VM7d(L)"        "AL-DC3(R)"        
#> [154] "AL-VP1l(R)"        "ATL(L)"            "lACA(R)"          
#> [157] "EBr3pw"            "VES(L)"            "MB(R)"            
#> [160] "PB(L1)"            "FBl2"              "AL-DL1(R)"        
#> [163] "AL(R)"             "AL-DA1(R)"         "SMP(L)"           
#> [166] "hemibrain"         "g1(R)"             "AL-DM4(L)"        
#> [169] "NO3(L)"            "LH(R)"             "AMMC"             
#> [172] "NO2(L)"            "SNP(L)"            "FBl4"             
#> [175] "AL-DA3(R)"         "AL-VM5d(R)"        "EB"               
#> [178] "LX(R)"             "g5(R)"             "AL-VA3(R)"        
#> [181] "g2(R)"             "PB(L7)"            "NO1(R)"           
#> [184] "PB(R1)"            "AL-DL3(R)"         "AL-VL2a(R)"       
#> [187] "GF(R)"             "mALT(R)"           "AL-VM7v(L)"       
#> [190] "b'L(L)"            "NO(L)"             "AL-DL5(R)"        
#> [193] "IB"                "RUB(L)"            "POC"              
#> [196] "EBr3am"            "GOR(L)"            "WED(R)"           
#> [199] "AL-DL4(L)"         "AL-DM3(L)"         "SPS(R)"           
#> [202] "PB(L6)"            "FBl5"              "PB(L5)"           
#> [205] "PB(L8)"            "PLP(R)"            "GOR(R)"           
#> [208] "AL-V(R)"           "SAD(-AMMC)"        "FBl9"             
#> [211] "AL-DA4l(R)"        "AL-DA4m(R)"        "PB(L2)"           
#> [214] "a2(R)"             "CRE(-RUB)(L)"      "AL-DC1(R)"        
#> [217] "AL-VA7l(R)"        "AL-VC2(R)"         "AL-VP3(R)"        
#> [220] "AL-DA2(L)"         "LO(R)"             "PB(R2)"           
#> [223] "AL-VA4(R)"         "ICL(L)"            "aL(L)"            
#> [226] "AL-VP1m(R)"        "SMP(R)"            "ROB(R)"           
#> [229] "AL-VA2(R)"         "AL-VP5(R)"         "AL-VP4(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "AB(R)"             "SCL(L)"            "PB(L7)"           
#>   [4] "PB(R4)"            "IPS(R)"            "FB"               
#>   [7] "PB(L9)"            "AOT(R)"            "PENP"             
#>  [10] "SIP(L)"            "g2(R)"             "LX(R)"            
#>  [13] "LO(R)"             "PB(L4)"            "AL(R)"            
#>  [16] "CRE(-ROB,-RUB)(R)" "b'2(R)"            "EBr3pw"           
#>  [19] "SNP(L)"            "FBl7"              "CA(R)"            
#>  [22] "NO2(R)"            "CX"                "NO"               
#>  [25] "lACA(R)"           "CRE(R)"            "SLP(R)"           
#>  [28] "AB(L)"             "ATL(L)"            "PB(R6)"           
#>  [31] "EBr3d"             "PB(R9)"            "g5(R)"            
#>  [34] "a'L(L)"            "mALT(L)"           "AMMC"             
#>  [37] "PB(L3)"            "PB(R7)"            "NO1(L)"           
#>  [40] "a'1(R)"            "a2(R)"             "GOR(L)"           
#>  [43] "mALT(R)"           "SNP(R)"            "EPA(L)"           
#>  [46] "GNG"               "bL(R)"             "POC"              
#>  [49] "hemibrain"         "RUB(R)"            "b1(R)"            
#>  [52] "WED(R)"            "EBr6"              "dACA(R)"          
#>  [55] "ME(R)"             "PVLP(R)"           "a1(R)"            
#>  [58] "PRW"               "VMNP"              "LH(R)"            
#>  [61] "b'L(L)"            "SPS(L)"            "GF(R)"            
#>  [64] "aL(R)"             "FBl2"              "RUB(L)"           
#>  [67] "LAL(L)"            "ICL(L)"            "MB(+ACA)(R)"      
#>  [70] "FBl1"              "INP"               "b'L(R)"           
#>  [73] "NO(R)"             "LAL(R)"            "g1(R)"            
#>  [76] "VES(L)"            "PB(R8)"            "ATL(R)"           
#>  [79] "a'3(R)"            "EB"                "PB(L1)"           
#>  [82] "FBl8"              "GOR(R)"            "NO(L)"            
#>  [85] "AOTU(R)"           "FBl4"              "IB"               
#>  [88] "SCL(R)"            "vACA(R)"           "gL(R)"            
#>  [91] "FBl9"              "AVLP(R)"           "aL(L)"            
#>  [94] "PB(R5)"            "b'1(R)"            "BU(L)"            
#>  [97] "NO3(L)"            "MB(R)"             "EBr2r4"           
#> [100] "PLP(R)"            "PB(R1)"            "CA(L)"            
#> [103] "MB(L)"             "SPS(R)"            "PED(R)"           
#> [106] "VES(R)"            "FBl6"              "VLNP(R)"          
#> [109] "SMP(L)"            "FBl5"              "g4(R)"            
#> [112] "EBr1"              "FLA(R)"            "LX(L)"            
#> [115] "PB(L2)"            "EBr3am"            "LOP(R)"           
#> [118] "LAL(-GA)(R)"       "a'L(R)"            "ROB(R)"           
#> [121] "AME(R)"            "NO2(L)"            "PB"               
#> [124] "FBl3"              "ICL(R)"            "NO3(R)"           
#> [127] "b2(R)"             "EBr5"              "a'2(R)"           
#> [130] "bL(L)"             "SAD(-AMMC)"        "PB(R3)"           
#> [133] "CAN(R)"            "GA(R)"             "OL(R)"            
#> [136] "GC"                "PB(L8)"            "NO1(R)"           
#> [139] "EPA(R)"            "PB(R2)"            "SMP(R)"           
#> [142] "PB(L5)"            "SAD"               "BU(R)"            
#> [145] "gL(L)"             "AL(L)"             "SIP(R)"           
#> [148] "PB(L6)"            "CRE(L)"            "a3(R)"            
#> [151] "g3(R)"            
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "aL(L)"             "AL-VC3l(R)"        "AL-VC2(R)"        
#>   [4] "LOP(R)"            "PB(L3)"            "AL-DA4m(R)"       
#>   [7] "BU(R)"             "a3(R)"             "b2(R)"            
#>  [10] "SIP(R)"            "NO2(R)"            "AL-VM4(R)"        
#>  [13] "AOT(R)"            "NO1(L)"            "b'L(R)"           
#>  [16] "ICL(R)"            "bL(R)"             "LAL(L)"           
#>  [19] "hemibrain"         "ME(R)"             "AL-DM2(L)"        
#>  [22] "AL-DM5(L)"         "AL-VC1(R)"         "g3(R)"            
#>  [25] "a2(R)"             "FB"                "SCL(R)"           
#>  [28] "AL-DA1(R)"         "b'1(R)"            "mALT(R)"          
#>  [31] "aL(R)"             "AL-DA3(R)"         "AL-DC3(R)"        
#>  [34] "CRE(-ROB,-RUB)(R)" "GC"                "LX(R)"            
#>  [37] "AL-DC1(L)"         "PB(R1)"            "FBl1"             
#>  [40] "ATL(L)"            "AME(R)"            "GOR(R)"           
#>  [43] "a'L(L)"            "AL-DM5(R)"         "AL-VA4(R)"        
#>  [46] "PB(L4)"            "AL-D(R)"           "AL-VA6(R)"        
#>  [49] "IPS(R)"            "b'L(L)"            "SIP(L)"           
#>  [52] "PB(R7)"            "AL-DL2d(R)"        "AL-VM7d(R)"       
#>  [55] "SPS(L)"            "AL-DA3(L)"         "NO(R)"            
#>  [58] "LO(R)"             "VLNP(R)"           "vACA(R)"          
#>  [61] "PB(L6)"            "PB(R3)"            "AL-VM3(R)"        
#>  [64] "SAD"               "BU(L)"             "SNP(L)"           
#>  [67] "AL-VC5(R)"         "AL-V(R)"           "mALT(L)"          
#>  [70] "AL-DA4m(L)"        "PVLP(R)"           "PB"               
#>  [73] "AL-D(L)"           "AL-DL2v(R)"        "CRE(-RUB)(L)"     
#>  [76] "AL-DA2(R)"         "AL-DM2(R)"         "AL-DP1l(R)"       
#>  [79] "EBr6"              "AL-VP2(R)"         "dACA(R)"          
#>  [82] "AL-DM4(R)"         "AL-VP1m(R)"        "PENP"             
#>  [85] "AL-DM3(L)"         "PB(R4)"            "EBr2r4"           
#>  [88] "PB(L9)"            "AL(R)"             "AL-VC3m(R)"       
#>  [91] "PLP(R)"            "NO3(L)"            "SAD(-AMMC)"       
#>  [94] "VES(R)"            "PB(L1)"            "AL-DL5(L)"        
#>  [97] "AL-VA6(L)"         "AL-DM3(R)"         "AL-DM4(L)"        
#> [100] "AL-DP1m(L)"        "EBr3pw"            "CA(R)"            
#> [103] "CA(L)"             "PRW"               "LH(R)"            
#> [106] "CAN(R)"            "INP"               "AL-VA1d(R)"       
#> [109] "AB(L)"             "EB"                "AL-DL4(R)"        
#> [112] "SNP(R)"            "RUB(L)"            "AL-DL5(R)"        
#> [115] "AL-VP3(R)"         "GF(R)"             "FB-column3"       
#> [118] "NO3(R)"            "AL-DM6(L)"         "PB(L7)"           
#> [121] "MB(+ACA)(R)"       "AL-DC2(L)"         "AL-VM7d(L)"       
#> [124] "ICL(L)"            "AL-VA7m(R)"        "PB(L5)"           
#> [127] "lACA(R)"           "FBl3"              "GNG"              
#> [130] "AL-DP1m(R)"        "SMP(R)"            "FLA(R)"           
#> [133] "PB(R8)"            "ATL(R)"            "AL-DA2(L)"        
#> [136] "PB(R2)"            "VES(L)"            "AL-DL1(R)"        
#> [139] "AL-VA1v(R)"        "AL-VP5(R)"         "AL-VL2a(R)"       
#> [142] "g2(R)"             "PB(L8)"            "AL-DL3(R)"        
#> [145] "bL(L)"             "LX(L)"             "SLP(R)"           
#> [148] "AL-DA4l(R)"        "g4(R)"             "AL-DL4(L)"        
#> [151] "CRE(L)"            "AL-VL2p(R)"        "AL-VM7v(R)"       
#> [154] "a'3(R)"            "AL(L)"             "FBl6"             
#> [157] "CRE(R)"            "AVLP(R)"           "PB(R6)"           
#> [160] "SPS(R)"            "NO1(R)"            "a'1(R)"           
#> [163] "MB(R)"             "AL-VC4(R)"         "POC"              
#> [166] "SCL(L)"            "EBr1"              "gL(L)"            
#> [169] "AL-VA5(R)"         "CX"                "g5(R)"            
#> [172] "FBl5"              "EPA(L)"            "AL-DM1(R)"        
#> [175] "AL-VM2(R)"         "EBr3am"            "LAL(-GA)(R)"      
#> [178] "EPA(R)"            "LAL(R)"            "OL(R)"            
#> [181] "g1(R)"             "a'L(R)"            "FBl4"             
#> [184] "NO"                "FBl2"              "b1(R)"            
#> [187] "GA(R)"             "NO(L)"             "WED(R)"           
#> [190] "AL-VM7v(L)"        "ROB(R)"            "FBl7"             
#> [193] "FBl9"              "AL-VP1d(R)"        "GOR(L)"           
#> [196] "gL(R)"             "AL-DC2(R)"         "AL-VA7l(R)"       
#> [199] "SMP(L)"            "MB(L)"             "AL-DC1(R)"        
#> [202] "AB(R)"             "AL-DC4(R)"         "RUB(R)"           
#> [205] "b'2(R)"            "EBr3d"             "AL-VM1(R)"        
#> [208] "AL-VP1l(R)"        "NO2(L)"            "AOTU(R)"          
#> [211] "PED(R)"            "VMNP"              "AL-VA2(R)"        
#> [214] "PB(R5)"            "PB(R9)"            "AL-DM6(R)"        
#> [217] "AL-VP4(R)"         "AL-VM5v(R)"        "PB(L2)"           
#> [220] "a1(R)"             "AL-DM1(L)"         "AL-VM5d(R)"       
#> [223] "IB"                "AMMC"              "EBr5"             
#> [226] "a'2(R)"            "AL-DC4(L)"         "FBl8"             
#> [229] "AL-VA3(R)"         "AL-VL1(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
#> $`hemibrain:v1.2`
#> $`hemibrain:v1.2`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2`$ROIs
#>   [1] "AL-VC1(R)"         "AL-VP4(R)"         "PB(L3)"           
#>   [4] "CRE(R)"            "AL-VA1v(R)"        "AL-VA6(R)"        
#>   [7] "g3(R)"             "EBr2r4"            "PVLP(R)"          
#>  [10] "AL-VM1(R)"         "EBr3am"            "PENP"             
#>  [13] "SMP(L)"            "b'1(R)"            "AL-DA3(L)"        
#>  [16] "AL-VM5d(R)"        "PB(R5)"            "CX"               
#>  [19] "b'L(R)"            "BU(L)"             "AL-DL3(R)"        
#>  [22] "gL(L)"             "a'1(R)"            "FB-column3"       
#>  [25] "FBl4"              "FBl5"              "PB(R2)"           
#>  [28] "ATL(R)"            "AL-DA4m(R)"        "NO3(R)"           
#>  [31] "NO1(R)"            "mALT(L)"           "AL-DM4(L)"        
#>  [34] "AL-VP1l(R)"        "AL-DL5(R)"         "AL-VM2(R)"        
#>  [37] "AMMC"              "g5(R)"             "FBl9"             
#>  [40] "GOR(R)"            "PB(R6)"            "b'2(R)"           
#>  [43] "GF(R)"             "PB(L5)"            "MB(R)"            
#>  [46] "RUB(L)"            "bL(R)"             "a3(R)"            
#>  [49] "AL-DP1m(R)"        "SCL(R)"            "dACA(R)"          
#>  [52] "b2(R)"             "OL(R)"             "ME(R)"            
#>  [55] "FBl2"              "PB(R4)"            "PB(R7)"           
#>  [58] "VMNP"              "AL-VM4(R)"         "CRE(L)"           
#>  [61] "AL-DC4(R)"         "AL-DM6(L)"         "NO1(L)"           
#>  [64] "AL-DC1(L)"         "INP"               "AL-V(R)"          
#>  [67] "BU(R)"             "NO"                "FBl1"             
#>  [70] "AL-DM1(R)"         "PRW"               "SIP(R)"           
#>  [73] "SNP(R)"            "CA(R)"             "AL-DA2(R)"        
#>  [76] "SPS(R)"            "EBr3d"             "ICL(L)"           
#>  [79] "AL-DC1(R)"         "AL-VC3m(R)"        "AOT(R)"           
#>  [82] "EBr5"              "LX(R)"             "AVLP(R)"          
#>  [85] "NO2(R)"            "AL-VA1d(R)"        "SIP(L)"           
#>  [88] "LO(R)"             "MB(L)"             "AL-D(L)"          
#>  [91] "FBl7"              "IPS(R)"            "AL-DL4(R)"        
#>  [94] "GOR(L)"            "LAL(-GA)(R)"       "AL-DC2(R)"        
#>  [97] "FB"                "AL-DC4(L)"         "AL-DL4(L)"        
#> [100] "AL-DA4l(R)"        "AL-VL2p(R)"        "a1(R)"            
#> [103] "GNG"               "AL-VA5(R)"         "AL-DL2d(R)"       
#> [106] "NO2(L)"            "SNP(L)"            "VES(L)"           
#> [109] "AL-DM3(R)"         "AL-VP1d(R)"        "AL-DM2(L)"        
#> [112] "CRE(-RUB)(L)"      "aL(R)"             "AL-DM5(L)"        
#> [115] "AL-VM7v(L)"        "LX(L)"             "PLP(R)"           
#> [118] "AL-DC2(L)"         "AL-DM1(L)"         "AL-VA6(L)"        
#> [121] "vACA(R)"           "AL-DA2(L)"         "AL-DL5(L)"        
#> [124] "ICL(R)"            "CAN(R)"            "LAL(L)"           
#> [127] "PB(R8)"            "AL-VL2a(R)"        "AL-VP1m(R)"       
#> [130] "SCL(L)"            "LAL(R)"            "AL-VP3(R)"        
#> [133] "AL-VC5(R)"         "PB(L8)"            "AL-DM2(R)"        
#> [136] "AL-DM3(L)"         "AL-VM3(R)"         "AL-DM4(R)"        
#> [139] "AL-VA3(R)"         "PB(L9)"            "aL(L)"            
#> [142] "a'L(L)"            "AL-VC2(R)"         "AL-VC3l(R)"       
#> [145] "LOP(R)"            "PB"                "bL(L)"            
#> [148] "lACA(R)"           "ROB(R)"            "b1(R)"            
#> [151] "AL-DM6(R)"         "AL-VA7l(R)"        "GA(R)"            
#> [154] "FLA(R)"            "AL-VL1(R)"         "SMP(R)"           
#> [157] "b'L(L)"            "AL-VA2(R)"         "LH(R)"            
#> [160] "NO(L)"             "PB(L6)"            "gL(R)"            
#> [163] "AL-VA4(R)"         "AME(R)"            "NO3(L)"           
#> [166] "FBl8"              "mALT(R)"           "AL-VM7v(R)"       
#> [169] "PB(L4)"            "AL-VM5v(R)"        "AL-VC4(R)"        
#> [172] "AB(L)"             "MB(+ACA)(R)"       "EPA(L)"           
#> [175] "AL-DP1l(R)"        "VLNP(R)"           "IB"               
#> [178] "SPS(L)"            "AL-DA4m(L)"        "EBr3pw"           
#> [181] "AL(R)"             "ATL(L)"            "hemibrain"        
#> [184] "RUB(R)"            "AL(L)"             "AL-DA3(R)"        
#> [187] "PB(L2)"            "WED(R)"            "PB(L1)"           
#> [190] "a'2(R)"            "g2(R)"             "EBr6"             
#> [193] "g4(R)"             "AL-DA1(R)"         "AL-DC3(R)"        
#> [196] "AL-DL1(R)"         "AL-DL2v(R)"        "GC"               
#> [199] "EBr1"              "POC"               "CRE(-ROB,-RUB)(R)"
#> [202] "a'L(R)"            "AL-VA7m(R)"        "g1(R)"            
#> [205] "AL-VP5(R)"         "AOTU(R)"           "SAD"              
#> [208] "PB(R9)"            "PB(R1)"            "AL-VM7d(L)"       
#> [211] "PB(L7)"            "SAD(-AMMC)"        "VES(R)"           
#> [214] "FBl3"              "AL-VP2(R)"         "PED(R)"           
#> [217] "AL-DP1m(L)"        "a'3(R)"            "AL-DM5(R)"        
#> [220] "AL-VM7d(R)"        "EB"                "EPA(R)"           
#> [223] "PB(R3)"            "AB(R)"             "SLP(R)"           
#> [226] "a2(R)"             "NO(R)"             "CA(L)"            
#> [229] "FBl6"              "AL-D(R)"          
#> 
#> $`hemibrain:v1.2`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                      
#> [1,] "GET"  "/*"                                      
#> [2,] "POST" "/api/v:ver/npexplorer/neuronmeta"        
#> [3,] "GET"  "/api/cached/dailytype"                   
#> [4,] "GET"  "/api/v:ver/npexplorer/commonconnectivity"
#> [5,] "GET"  "/api/npexplorer/distribution"            
#> [6,] "POST" "/api/v:ver/raw/cypher/cypher"            
#>      name                                                                                        
#> [1,] "github.com/labstack/echo.common.static.func1"                                              
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/cached.cypherAPI.getDailyType-fm"          
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }