Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "g3(R)"             "b'L(R)"            "LX(R)"            
#>   [4] "g4(R)"             "EBr3pw"            "ATL(R)"           
#>   [7] "SPS(R)"            "PB(R1)"            "PB(L9)"           
#>  [10] "LO(R)"             "ICL(R)"            "AL(R)"            
#>  [13] "LOP(R)"            "PB(R7)"            "PLP(R)"           
#>  [16] "SCL(L)"            "a1(R)"             "EBr5"             
#>  [19] "PB(R8)"            "BU(R)"             "FBl4"             
#>  [22] "PB(L8)"            "EBr3am"            "SLP(R)"           
#>  [25] "PED(R)"            "VMNP"              "SNP(R)"           
#>  [28] "AL(L)"             "FBl7"              "PB(R3)"           
#>  [31] "a3(R)"             "MB(R)"             "AOTU(R)"          
#>  [34] "CRE(R)"            "FLA(R)"            "g5(R)"            
#>  [37] "PB(L3)"            "GC"                "EBr3d"            
#>  [40] "SIP(R)"            "NO3(L)"            "SAD"              
#>  [43] "MB(+ACA)(R)"       "aL(L)"             "RUB(L)"           
#>  [46] "bL(R)"             "mALT(R)"           "aL(R)"            
#>  [49] "NO3(R)"            "EBr6"              "NO1(R)"           
#>  [52] "SPS(L)"            "ROB(R)"            "a'L(L)"           
#>  [55] "PB(R2)"            "CRE(-ROB,-RUB)(R)" "ATL(L)"           
#>  [58] "g2(R)"             "PB(R9)"            "AVLP(R)"          
#>  [61] "b'L(L)"            "PB"                "PB(L2)"           
#>  [64] "lACA(R)"           "g1(R)"             "EBr1"             
#>  [67] "FBl6"              "FBl9"              "RUB(R)"           
#>  [70] "PB(L5)"            "hemibrain"         "bL(L)"            
#>  [73] "PVLP(R)"           "CA(R)"             "PB(L4)"           
#>  [76] "IB"                "dACA(R)"           "INP"              
#>  [79] "NO2(L)"            "PB(R6)"            "SCL(R)"           
#>  [82] "ME(R)"             "NO(R)"             "ICL(L)"           
#>  [85] "EBr2r4"            "gL(R)"             "SAD(-AMMC)"       
#>  [88] "LH(R)"             "AME(R)"            "LAL(-GA)(R)"      
#>  [91] "GNG"               "VES(L)"            "PRW"              
#>  [94] "LX(L)"             "a'3(R)"            "GF(R)"            
#>  [97] "VES(R)"            "AMMC"              "SIP(L)"           
#> [100] "FB"                "a'L(R)"            "NO"               
#> [103] "a'1(R)"            "IPS(R)"            "OL(R)"            
#> [106] "CX"                "b'2(R)"            "EPA(R)"           
#> [109] "LAL(R)"            "mALT(L)"           "PENP"             
#> [112] "PB(L6)"            "FBl8"              "NO(L)"            
#> [115] "POC"               "CRE(L)"            "FBl2"             
#> [118] "b1(R)"             "EPA(L)"            "CA(L)"            
#> [121] "GOR(R)"            "LAL(L)"            "GA(R)"            
#> [124] "NO1(L)"            "b2(R)"             "EB"               
#> [127] "FBl3"              "AB(R)"             "a2(R)"            
#> [130] "gL(L)"             "PB(R4)"            "SMP(R)"           
#> [133] "WED(R)"            "MB(L)"             "PB(L1)"           
#> [136] "PB(L7)"            "SNP(L)"            "AOT(R)"           
#> [139] "AB(L)"             "a'2(R)"            "FBl1"             
#> [142] "CAN(R)"            "vACA(R)"           "b'1(R)"           
#> [145] "SMP(L)"            "GOR(L)"            "BU(L)"            
#> [148] "FBl5"              "NO2(R)"            "VLNP(R)"          
#> [151] "PB(R5)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method    path                                     
#> [1,] "CONNECT" "/api/*"                                 
#> [2,] "GET"     "/api/v:ver/cached/roicompleteness"      
#> [3,] "GET"     "/api/v:ver/npexplorer/neuronmetavals"   
#> [4,] "POST"    "/api/v:ver/npexplorer/rankedtable"      
#> [5,] "POST"    "/api/npexplorer/celltype/:dataset/:type"
#> [6,] "REPORT"  "/api"                                   
#>      name                                                                                    
#> [1,] "github.com/labstack/echo.glob..func1"                                                  
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCellType-fm"
#> [6,] "github.com/labstack/echo.glob..func1"                                                  

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }