Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "b'2(R)"            "FB"                "gL(L)"            
#>   [4] "AL(L)"             "SNP(L)"            "AMMC"             
#>   [7] "SMP(L)"            "PED(R)"            "GOR(L)"           
#>  [10] "dACA(R)"           "FBl8"              "GOR(R)"           
#>  [13] "PB(R5)"            "CA(L)"             "GF(R)"            
#>  [16] "NO3(R)"            "SMP(R)"            "a1(R)"            
#>  [19] "NO1(R)"            "SPS(L)"            "VMNP"             
#>  [22] "CRE(L)"            "AB(R)"             "BU(R)"            
#>  [25] "EPA(R)"            "SCL(L)"            "PB(R8)"           
#>  [28] "b'1(R)"            "EBr3am"            "vACA(R)"          
#>  [31] "CX"                "SIP(L)"            "PB(L6)"           
#>  [34] "ICL(L)"            "MB(+ACA)(R)"       "g3(R)"            
#>  [37] "aL(R)"             "b2(R)"             "BU(L)"            
#>  [40] "SIP(R)"            "ROB(R)"            "SAD(-AMMC)"       
#>  [43] "lACA(R)"           "PB(R7)"            "FLA(R)"           
#>  [46] "PB(L8)"            "EB"                "a'1(R)"           
#>  [49] "CAN(R)"            "a3(R)"             "SLP(R)"           
#>  [52] "RUB(R)"            "hemibrain"         "CRE(R)"           
#>  [55] "PB(L9)"            "PB(R6)"            "WED(R)"           
#>  [58] "PB(L7)"            "FBl7"              "CA(R)"            
#>  [61] "NO2(R)"            "LOP(R)"            "a'3(R)"           
#>  [64] "ATL(L)"            "AOTU(R)"           "PB(R4)"           
#>  [67] "PVLP(R)"           "NO1(L)"            "NO"               
#>  [70] "FBl1"              "SAD"               "g4(R)"            
#>  [73] "EBr3pw"            "FBl4"              "aL(L)"            
#>  [76] "PB(L1)"            "VES(R)"            "PB(R2)"           
#>  [79] "LAL(R)"            "GC"                "PB(R1)"           
#>  [82] "LH(R)"             "IB"                "SPS(R)"           
#>  [85] "PLP(R)"            "EBr2r4"            "b1(R)"            
#>  [88] "CRE(-ROB,-RUB)(R)" "PB(L3)"            "EBr6"             
#>  [91] "FBl3"              "AVLP(R)"           "PB(L4)"           
#>  [94] "a'L(L)"            "GA(R)"             "g1(R)"            
#>  [97] "bL(L)"             "RUB(L)"            "b'L(L)"           
#> [100] "VES(L)"            "gL(R)"             "FBl5"             
#> [103] "AB(L)"             "g2(R)"             "SNP(R)"           
#> [106] "NO(R)"             "FBl9"              "MB(L)"            
#> [109] "g5(R)"             "LO(R)"             "mALT(L)"          
#> [112] "NO2(L)"            "b'L(R)"            "SCL(R)"           
#> [115] "ME(R)"             "a'L(R)"            "FBl6"             
#> [118] "ICL(R)"            "bL(R)"             "PB(L5)"           
#> [121] "PENP"              "LX(R)"             "FBl2"             
#> [124] "LX(L)"             "INP"               "AL(R)"            
#> [127] "POC"               "a2(R)"             "EBr5"             
#> [130] "NO(L)"             "LAL(L)"            "AME(R)"           
#> [133] "OL(R)"             "LAL(-GA)(R)"       "PB"               
#> [136] "EBr3d"             "IPS(R)"            "AOT(R)"           
#> [139] "EBr1"              "PB(R9)"            "ATL(R)"           
#> [142] "PB(R3)"            "PB(L2)"            "PRW"              
#> [145] "mALT(R)"           "a'2(R)"            "EPA(L)"           
#> [148] "GNG"               "MB(R)"             "NO3(L)"           
#> [151] "VLNP(R)"          
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                    
#> [1,] "GET"  "/api/v:ver/dbmeta/database"            
#> [2,] "POST" "/api/raw/cypher/transaction/:id/commit"
#> [3,] "GET"  "/api/help/*"                           
#> [4,] "GET"  "/api/v:ver/npexplorer/autapses"        
#> [5,] "POST" "/api/v:ver/npexplorer/distribution"    
#> [6,] "POST" "/api/npexplorer/findneurons"           
#>      name                                                                                       
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [2,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                   
#> [3,] "github.com/labstack/echo.common.static.func1"                                             
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getFindNeurons-fm"

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }