Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "SPS(R)"            "g3(R)"             "LAL(-GA)(R)"      
#>   [4] "NO1(L)"            "b2(R)"             "EBr1"             
#>   [7] "AL-DA4m(L)"        "AL-DA4m(R)"        "mALT(R)"          
#>  [10] "WED(R)"            "BU(L)"             "vACA(R)"          
#>  [13] "AMMC"              "SIP(L)"            "lACA(R)"          
#>  [16] "a'1(R)"            "FBl3"              "AL-DM5(R)"        
#>  [19] "AL-V(R)"           "POC"               "GNG"              
#>  [22] "NO2(L)"            "LX(R)"             "EPA(R)"           
#>  [25] "bL(R)"             "AL-DM3(R)"         "AL-DL4(R)"        
#>  [28] "a'L(R)"            "FBl6"              "AL-DM2(R)"        
#>  [31] "AL-VC3m(R)"        "AL-VP5(R)"         "MB(L)"            
#>  [34] "BU(R)"             "ATL(L)"            "PB(R8)"           
#>  [37] "AME(R)"            "AL-DM1(R)"         "AL-VL1(R)"        
#>  [40] "FB-column3"        "IB"                "SCL(L)"           
#>  [43] "SMP(L)"            "AL-DP1m(L)"        "MB(+ACA)(R)"      
#>  [46] "AL-VM4(R)"         "OL(R)"             "VES(L)"           
#>  [49] "AL-DC2(R)"         "GOR(L)"            "EBr2r4"           
#>  [52] "AL(L)"             "ROB(R)"            "FBl1"             
#>  [55] "PB(R4)"            "AL-VM3(R)"         "AL-DM2(L)"        
#>  [58] "VES(R)"            "AL-VA4(R)"         "NO3(R)"           
#>  [61] "FB"                "AL-DA2(R)"         "AL-DP1m(R)"       
#>  [64] "RUB(R)"            "PB(L3)"            "PB"               
#>  [67] "NO3(L)"            "ATL(R)"            "RUB(L)"           
#>  [70] "AL-VC2(R)"         "EBr3d"             "NO"               
#>  [73] "PB(L6)"            "ICL(R)"            "AL-DL1(R)"        
#>  [76] "AL-VM7d(R)"        "AB(R)"             "AL-DC3(R)"        
#>  [79] "AL-DC4(R)"         "AL-DM4(L)"         "SLP(R)"           
#>  [82] "AL-VC5(R)"         "PB(L5)"            "gL(L)"            
#>  [85] "PB(R2)"            "GOR(R)"            "AL-VM1(R)"        
#>  [88] "PB(L1)"            "PVLP(R)"           "IPS(R)"           
#>  [91] "AL-DL5(R)"         "VMNP"              "FBl7"             
#>  [94] "GF(R)"             "FBl9"              "AL-VP1l(R)"       
#>  [97] "b'1(R)"            "b'L(R)"            "SNP(R)"           
#> [100] "EB"                "EPA(L)"            "VLNP(R)"          
#> [103] "g5(R)"             "SAD(-AMMC)"        "AL-DM6(R)"        
#> [106] "FLA(R)"            "AL-D(L)"           "AL-VA7m(R)"       
#> [109] "b'2(R)"            "AL-DM6(L)"         "NO(L)"            
#> [112] "AL-VM5v(R)"        "gL(R)"             "MB(R)"            
#> [115] "AL-VA1v(R)"        "LH(R)"             "a1(R)"            
#> [118] "PENP"              "LAL(L)"            "g1(R)"            
#> [121] "SCL(R)"            "PLP(R)"            "EBr3pw"           
#> [124] "AL-DL3(R)"         "EBr5"              "NO(R)"            
#> [127] "AL-DA4l(R)"        "AL-DC1(R)"         "CA(L)"            
#> [130] "PB(R1)"            "AVLP(R)"           "NO2(R)"           
#> [133] "aL(L)"             "AL-VP4(R)"         "AL-DM1(L)"        
#> [136] "AL-VC1(R)"         "AOT(R)"            "SAD"              
#> [139] "AL-DA3(L)"         "PB(L7)"            "LX(L)"            
#> [142] "AL-VM2(R)"         "GA(R)"             "AL-VA6(L)"        
#> [145] "AL-VM7v(L)"        "LO(R)"             "CAN(R)"           
#> [148] "a'L(L)"            "ME(R)"             "AL-VM7d(L)"       
#> [151] "a3(R)"             "AL-VM5d(R)"        "a'2(R)"           
#> [154] "FBl2"              "b'L(L)"            "PB(L4)"           
#> [157] "AL-DA3(R)"         "aL(R)"             "AL-DL4(L)"        
#> [160] "SMP(R)"            "AOTU(R)"           "PB(L8)"           
#> [163] "AL-DA2(L)"         "AL-DC1(L)"         "PB(L9)"           
#> [166] "AL-DA1(R)"         "AL-DP1l(R)"        "AL-VA5(R)"        
#> [169] "PB(R6)"            "hemibrain"         "g4(R)"            
#> [172] "SNP(L)"            "AL-VP1d(R)"        "mALT(L)"          
#> [175] "AL-DC2(L)"         "AL-D(R)"           "AL-VA3(R)"        
#> [178] "a'3(R)"            "AL-DM3(L)"         "IB112"            
#> [181] "FBl8"              "AL-VC4(R)"         "AL-VM7v(R)"       
#> [184] "PB(R5)"            "PRW"               "CRE(L)"           
#> [187] "FBl4"              "AB(L)"             "CX"               
#> [190] "EBr6"              "SPS(L)"            "AL-VP3(R)"        
#> [193] "PB(L2)"            "INP"               "LAL(R)"           
#> [196] "AL-VA2(R)"         "AL-VL2a(R)"        "g2(R)"            
#> [199] "dACA(R)"           "PB(R9)"            "a2(R)"            
#> [202] "AL-VA6(R)"         "LOP(R)"            "AL-DM5(L)"        
#> [205] "SIP(R)"            "b1(R)"             "AL(R)"            
#> [208] "FBl5"              "AL-VP2(R)"         "CRE(R)"           
#> [211] "AL-DL2v(R)"        "CRE(-RUB)(L)"      "CA(R)"            
#> [214] "NO1(R)"            "bL(L)"             "AL-VA7l(R)"       
#> [217] "EBr3am"            "AL-DL2d(R)"        "AL-VC3l(R)"       
#> [220] "GC"                "AL-DM4(R)"         "PED(R)"           
#> [223] "AL-DC4(L)"         "AL-DL5(L)"         "ICL(L)"           
#> [226] "AL-VL2p(R)"        "PB(R7)"            "PB(R3)"           
#> [229] "CRE(-ROB,-RUB)(R)" "AL-VP1m(R)"        "AL-VA1d(R)"       
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "BU(L)"             "CRE(L)"            "PB(R4)"           
#>   [4] "PB(L5)"            "SNP(R)"            "g4(R)"            
#>   [7] "gL(R)"             "FBl1"              "NO(L)"            
#>  [10] "GOR(L)"            "NO(R)"             "bL(L)"            
#>  [13] "ICL(L)"            "g5(R)"             "NO3(R)"           
#>  [16] "SCL(R)"            "PB(L9)"            "LX(L)"            
#>  [19] "NO2(L)"            "PENP"              "lACA(R)"          
#>  [22] "CAN(R)"            "POC"               "b'1(R)"           
#>  [25] "PB(L3)"            "PED(R)"            "LO(R)"            
#>  [28] "NO2(R)"            "GOR(R)"            "PB(R5)"           
#>  [31] "dACA(R)"           "PB(L8)"            "EBr2r4"           
#>  [34] "PVLP(R)"           "MB(+ACA)(R)"       "FBl2"             
#>  [37] "NO1(L)"            "PB(L1)"            "ME(R)"            
#>  [40] "BU(R)"             "FBl3"              "GNG"              
#>  [43] "CRE(-ROB,-RUB)(R)" "FLA(R)"            "LX(R)"            
#>  [46] "EBr3d"             "a'1(R)"            "IPS(R)"           
#>  [49] "CX"                "a1(R)"             "PB"               
#>  [52] "AL(L)"             "CA(R)"             "ICL(R)"           
#>  [55] "MB(L)"             "b'L(L)"            "AL(R)"            
#>  [58] "EPA(L)"            "LAL(R)"            "a2(R)"            
#>  [61] "PB(R7)"            "ROB(R)"            "ATL(R)"           
#>  [64] "INP"               "a'L(L)"            "vACA(R)"          
#>  [67] "AOTU(R)"           "SPS(L)"            "LH(R)"            
#>  [70] "PB(R3)"            "a3(R)"             "GC"               
#>  [73] "g2(R)"             "a'2(R)"            "PB(R1)"           
#>  [76] "ATL(L)"            "CRE(R)"            "VES(R)"           
#>  [79] "PB(R6)"            "GF(R)"             "FB"               
#>  [82] "b'L(R)"            "SNP(L)"            "FBl8"             
#>  [85] "aL(L)"             "LAL(-GA)(R)"       "SMP(R)"           
#>  [88] "PB(L2)"            "g3(R)"             "AVLP(R)"          
#>  [91] "MB(R)"             "VLNP(R)"           "GA(R)"            
#>  [94] "aL(R)"             "NO1(R)"            "gL(L)"            
#>  [97] "FBl7"              "FBl9"              "SAD(-AMMC)"       
#> [100] "bL(R)"             "PB(L4)"            "AOT(R)"           
#> [103] "b2(R)"             "AB(R)"             "EBr5"             
#> [106] "PB(R9)"            "EBr3pw"            "FBl5"             
#> [109] "AME(R)"            "EBr3am"            "hemibrain"        
#> [112] "SLP(R)"            "SCL(L)"            "AMMC"             
#> [115] "LAL(L)"            "EB"                "OL(R)"            
#> [118] "RUB(R)"            "a'L(R)"            "VES(L)"           
#> [121] "IB"                "SMP(L)"            "SPS(R)"           
#> [124] "RUB(L)"            "AB(L)"             "SAD"              
#> [127] "PRW"               "EBr6"              "PLP(R)"           
#> [130] "PB(L7)"            "FBl4"              "b'2(R)"           
#> [133] "g1(R)"             "EBr1"              "SIP(R)"           
#> [136] "PB(L6)"            "PB(R2)"            "b1(R)"            
#> [139] "mALT(L)"           "SIP(L)"            "mALT(R)"          
#> [142] "WED(R)"            "VMNP"              "EPA(R)"           
#> [145] "NO"                "CA(L)"             "PB(R8)"           
#> [148] "NO3(L)"            "FBl6"              "LOP(R)"           
#> [151] "a'3(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "VLNP(R)"           "b'L(R)"            "PB"               
#>   [4] "AL-DM5(L)"         "AL-VP5(R)"         "LH(R)"            
#>   [7] "AL-DA3(R)"         "AL-VP1d(R)"        "AOT(R)"           
#>  [10] "PB(R7)"            "PB(L7)"            "MB(+ACA)(R)"      
#>  [13] "FBl4"              "FBl1"              "AL-DA2(R)"        
#>  [16] "AL-DL2v(R)"        "b'1(R)"            "gL(L)"            
#>  [19] "EBr3pw"            "vACA(R)"           "FBl3"             
#>  [22] "FBl5"              "bL(R)"             "AL-DA4m(R)"       
#>  [25] "CA(R)"             "NO(L)"             "AL-VP1m(R)"       
#>  [28] "a'3(R)"            "g2(R)"             "POC"              
#>  [31] "PB(L3)"            "PB(R8)"            "GOR(R)"           
#>  [34] "RUB(L)"            "b'2(R)"            "SPS(L)"           
#>  [37] "EBr2r4"            "AL-D(L)"           "PB(R1)"           
#>  [40] "PB(R3)"            "AL-VP4(R)"         "AL-DM2(L)"        
#>  [43] "SIP(R)"            "AME(R)"            "LX(L)"            
#>  [46] "PB(L2)"            "aL(R)"             "AL-DA4l(R)"       
#>  [49] "AL-DC1(R)"         "AL-DC3(R)"         "b2(R)"            
#>  [52] "NO1(R)"            "CA(L)"             "LAL(L)"           
#>  [55] "BU(L)"             "AL-DM2(R)"         "PB(L5)"           
#>  [58] "EBr6"              "AL-DC2(L)"         "AL-DC4(R)"        
#>  [61] "AL-VA7l(R)"        "AL-VC3m(R)"        "AL-VM5d(R)"       
#>  [64] "GA(R)"             "VES(L)"            "AL-VC5(R)"        
#>  [67] "mALT(L)"           "VMNP"              "BU(R)"            
#>  [70] "PED(R)"            "AL-DP1m(L)"        "AL-VM2(R)"        
#>  [73] "SMP(R)"            "EB"                "AL-DL3(R)"        
#>  [76] "AL-VA1d(R)"        "a'2(R)"            "INP"              
#>  [79] "MB(R)"             "AL-DM3(R)"         "AL-DA3(L)"        
#>  [82] "AL-DM6(L)"         "g5(R)"             "PB(R2)"           
#>  [85] "SPS(R)"            "CRE(R)"            "NO(R)"            
#>  [88] "AL-VM4(R)"         "g4(R)"             "NO"               
#>  [91] "ROB(R)"            "AL-DL5(R)"         "ME(R)"            
#>  [94] "AL-DA4m(L)"        "AL-VA6(L)"         "b'L(L)"           
#>  [97] "a1(R)"             "SMP(L)"            "EBr1"             
#> [100] "AL-DM4(L)"         "PB(R9)"            "SAD(-AMMC)"       
#> [103] "FB-column3"        "LAL(R)"            "FBl2"             
#> [106] "ATL(R)"            "AL-VL2a(R)"        "LX(R)"            
#> [109] "NO3(L)"            "AL-VP2(R)"         "SAD"              
#> [112] "SNP(R)"            "gL(R)"             "PRW"              
#> [115] "AL(L)"             "AL-DC4(L)"         "a'L(L)"           
#> [118] "IB"                "GF(R)"             "EBr5"             
#> [121] "FLA(R)"            "EPA(L)"            "AL-DM5(R)"        
#> [124] "AL-VA3(R)"         "NO3(R)"            "MB(L)"            
#> [127] "AB(R)"             "IPS(R)"            "CRE(-ROB,-RUB)(R)"
#> [130] "g3(R)"             "VES(R)"            "AL(R)"            
#> [133] "AL-VA2(R)"         "AL-VA1v(R)"        "AL-VP1l(R)"       
#> [136] "EPA(R)"            "AL-DC1(L)"         "a'1(R)"           
#> [139] "FBl8"              "LAL(-GA)(R)"       "PLP(R)"           
#> [142] "NO2(R)"            "dACA(R)"           "AL-DA2(L)"        
#> [145] "CRE(-RUB)(L)"      "AL-VC3l(R)"        "AL-VM7v(L)"       
#> [148] "AL-VA5(R)"         "AL-VP3(R)"         "AMMC"             
#> [151] "SCL(L)"            "g1(R)"             "PB(L1)"           
#> [154] "ICL(L)"            "AL-VL2p(R)"        "OL(R)"            
#> [157] "AL-DM1(R)"         "AL-DP1m(R)"        "CX"               
#> [160] "SIP(L)"            "SLP(R)"            "AL-DC2(R)"        
#> [163] "AL-VL1(R)"         "FBl6"              "AL-DP1l(R)"       
#> [166] "hemibrain"         "AL-DM1(L)"         "CRE(L)"           
#> [169] "LO(R)"             "AOTU(R)"           "CAN(R)"           
#> [172] "EBr3am"            "ATL(L)"            "mALT(R)"          
#> [175] "a'L(R)"            "AL-DM3(L)"         "AL-VA7m(R)"       
#> [178] "AL-VA6(R)"         "AL-DL4(L)"         "AL-DL4(R)"        
#> [181] "AL-DM6(R)"         "AL-VA4(R)"         "FB"               
#> [184] "PB(L4)"            "AL-D(R)"           "WED(R)"           
#> [187] "PB(L8)"            "PB(R4)"            "GNG"              
#> [190] "PB(L6)"            "PVLP(R)"           "PB(R5)"           
#> [193] "NO1(L)"            "AL-DA1(R)"         "AB(L)"            
#> [196] "AL-VC4(R)"         "NO2(L)"            "b1(R)"            
#> [199] "LOP(R)"            "PB(R6)"            "GC"               
#> [202] "FBl9"              "ICL(R)"            "RUB(R)"           
#> [205] "AL-VM7v(R)"        "a2(R)"             "SCL(R)"           
#> [208] "AL-VM7d(L)"        "AL-DL1(R)"         "AL-DL2d(R)"       
#> [211] "AL-DM4(R)"         "AL-VC2(R)"         "PB(L9)"           
#> [214] "AL-VM5v(R)"        "AL-VM7d(R)"        "AL-V(R)"          
#> [217] "AL-DL5(L)"         "FBl7"              "EBr3d"            
#> [220] "a3(R)"             "PENP"              "SNP(L)"           
#> [223] "AL-VC1(R)"         "AL-VM3(R)"         "bL(L)"            
#> [226] "aL(L)"             "lACA(R)"           "GOR(L)"           
#> [229] "AVLP(R)"           "AL-VM1(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "PB(R6)"            "ICL(L)"            "AL-DM3(R)"        
#>   [4] "LO(R)"             "AL-VL2p(R)"        "AL-VM4(R)"        
#>   [7] "AL-VC4(R)"         "AL-VM3(R)"         "CRE(R)"           
#>  [10] "g4(R)"             "FBl2"              "AL-DL4(R)"        
#>  [13] "g5(R)"             "FBl1"              "b1(R)"            
#>  [16] "AL-DC1(R)"         "EB"                "AL-VA3(R)"        
#>  [19] "ATL(L)"            "SMP(L)"            "PRW"              
#>  [22] "SPS(L)"            "PB(L7)"            "IB"               
#>  [25] "AL-DA3(L)"         "ICL(R)"            "a'L(L)"           
#>  [28] "RUB(R)"            "AL-DM2(R)"         "AL-VA7m(R)"       
#>  [31] "AL-DM4(L)"         "SAD(-AMMC)"        "VES(R)"           
#>  [34] "LAL(R)"            "PENP"              "AB(L)"            
#>  [37] "AL-DL5(L)"         "AL-VM7d(L)"        "SNP(L)"           
#>  [40] "AL-D(R)"           "CX"                "a2(R)"            
#>  [43] "g2(R)"             "GOR(R)"            "AL-VA2(R)"        
#>  [46] "aL(R)"             "PB(L1)"            "ME(R)"            
#>  [49] "a'L(R)"            "PVLP(R)"           "PB(R8)"           
#>  [52] "AL-VM5v(R)"        "WED(R)"            "EBr5"             
#>  [55] "AL-VA4(R)"         "SIP(L)"            "LX(L)"            
#>  [58] "AL-DL1(R)"         "VLNP(R)"           "NO(R)"            
#>  [61] "NO1(L)"            "dACA(R)"           "gL(L)"            
#>  [64] "SIP(R)"            "AL-VP4(R)"         "b'1(R)"           
#>  [67] "EBr1"              "AL(L)"             "AL-DL4(L)"        
#>  [70] "AL-DP1l(R)"        "PB(R9)"            "AL-VM2(R)"        
#>  [73] "hemibrain"         "lACA(R)"           "AL-DL5(R)"        
#>  [76] "AL-VC2(R)"         "MB(+ACA)(R)"       "GNG"              
#>  [79] "AL-DM4(R)"         "AL-DM5(R)"         "PB(L3)"           
#>  [82] "AL-DM1(R)"         "AL-VC3m(R)"        "b2(R)"            
#>  [85] "FBl4"              "ATL(R)"            "AL-VP1m(R)"       
#>  [88] "PB(L5)"            "b'2(R)"            "PB(L8)"           
#>  [91] "AL-DC3(R)"         "EBr3am"            "AL-DA2(R)"        
#>  [94] "AOT(R)"            "SNP(R)"            "AL-DM2(L)"        
#>  [97] "EBr3pw"            "PB(R3)"            "LX(R)"            
#> [100] "PB(R2)"            "NO2(R)"            "PB"               
#> [103] "PB(L4)"            "AL-VA7l(R)"        "AL-VP2(R)"        
#> [106] "LOP(R)"            "GC"                "AL-DA2(L)"        
#> [109] "AL-VM7v(L)"        "CAN(R)"            "SLP(R)"           
#> [112] "EBr2r4"            "ROB(R)"            "PB(R1)"           
#> [115] "INP"               "gL(R)"             "LH(R)"            
#> [118] "AL-DL2v(R)"        "FB"                "AL-D(L)"          
#> [121] "FBl3"              "FBl9"              "SPS(R)"           
#> [124] "AVLP(R)"           "AL-DC4(R)"         "AL-VP5(R)"        
#> [127] "LAL(L)"            "NO"                "PB(R4)"           
#> [130] "AL-VP1l(R)"        "EBr6"              "NO3(L)"           
#> [133] "FB-column3"        "AL-VM5d(R)"        "PLP(R)"           
#> [136] "EPA(R)"            "bL(L)"             "AL-VC5(R)"        
#> [139] "AL-DC2(R)"         "AL-DM6(R)"         "AL-VC3l(R)"       
#> [142] "AL-VM7v(R)"        "SAD"               "GOR(L)"           
#> [145] "AL-DA4m(L)"        "MB(R)"             "PED(R)"           
#> [148] "AL-VP1d(R)"        "POC"               "PB(R5)"           
#> [151] "GF(R)"             "a1(R)"             "AL-DA1(R)"        
#> [154] "AL-DL3(R)"         "vACA(R)"           "CRE(-ROB,-RUB)(R)"
#> [157] "AL-VM1(R)"         "FLA(R)"            "BU(R)"            
#> [160] "AL-DP1m(R)"        "FBl7"              "NO(L)"            
#> [163] "AL-DM1(L)"         "CRE(L)"            "mALT(L)"          
#> [166] "AL-VA6(L)"         "FBl8"              "b'L(L)"           
#> [169] "AL-DL2d(R)"        "AL-V(R)"           "a'3(R)"           
#> [172] "GA(R)"             "BU(L)"             "AL-DM3(L)"        
#> [175] "AL-VL2a(R)"        "mALT(R)"           "SCL(R)"           
#> [178] "AL-DA4m(R)"        "a'2(R)"            "PB(L9)"           
#> [181] "AOTU(R)"           "FBl5"              "AL-DA3(R)"        
#> [184] "MB(L)"             "AB(R)"             "AL-VM7d(R)"       
#> [187] "IPS(R)"            "bL(R)"             "VES(L)"           
#> [190] "AL-DC2(L)"         "AL-DM5(L)"         "PB(L6)"           
#> [193] "a'1(R)"            "OL(R)"             "AL(R)"            
#> [196] "AL-VA5(R)"         "AL-VL1(R)"         "CA(L)"            
#> [199] "PB(R7)"            "FBl6"              "RUB(L)"           
#> [202] "AL-VC1(R)"         "aL(L)"             "PB(L2)"           
#> [205] "NO1(R)"            "VMNP"              "AL-DC4(L)"        
#> [208] "AL-DP1m(L)"        "AL-VP3(R)"         "g3(R)"            
#> [211] "g1(R)"             "AL-DM6(L)"         "AL-DA4l(R)"       
#> [214] "NO2(L)"            "LAL(-GA)(R)"       "AL-DC1(L)"        
#> [217] "AME(R)"            "AL-VA1v(R)"        "AL-VA6(R)"        
#> [220] "AMMC"              "NO3(R)"            "b'L(R)"           
#> [223] "EBr3d"             "CA(R)"             "SMP(R)"           
#> [226] "SCL(L)"            "CRE(-RUB)(L)"      "EPA(L)"           
#> [229] "a3(R)"             "AL-VA1d(R)"       
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method    path                               
#> [1,] "GET"     "/api/v:ver/dbmeta/database"       
#> [2,] "GET"     "/api/npexplorer/findneurons"      
#> [3,] "POST"    "/api/npexplorer/distribution"     
#> [4,] "POST"    "/api/roimeshes/mesh/:dataset/:roi"
#> [5,] "CONNECT" "/api/*"                           
#> [6,] "GET"     "/api/v:ver/dbmeta/version"        
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getFindNeurons-fm" 
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
#> [5,] "github.com/labstack/echo/v4.glob..func1"                                                   
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }