Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "b1(R)"             "a'L(L)"            "AL-VM4(R)"        
#>   [4] "AL-VM7v(R)"        "RUB(R)"            "CX"               
#>   [7] "AL-DM5(L)"         "ICL(R)"            "LX(L)"            
#>  [10] "AL-DL5(R)"         "PLP(R)"            "bL(R)"            
#>  [13] "a'3(R)"            "AOTU(R)"           "LAL(-GA)(R)"      
#>  [16] "GC"                "ROB(R)"            "AL-DA2(R)"        
#>  [19] "lACA(R)"           "LH(R)"             "hemibrain"        
#>  [22] "gL(L)"             "FBl2"              "POC"              
#>  [25] "NO3(L)"            "AL-DL3(R)"         "BU(L)"            
#>  [28] "MB(R)"             "LOP(R)"            "PED(R)"           
#>  [31] "CA(R)"             "AL-DP1m(R)"        "NO2(L)"           
#>  [34] "b'L(L)"            "AL-DC4(R)"         "AL-VP3(R)"        
#>  [37] "aL(L)"             "AL-DC1(R)"         "PB(L3)"           
#>  [40] "NO1(R)"            "AL-D(L)"           "SIP(R)"           
#>  [43] "PB(R1)"            "bL(L)"             "AL-VM7d(R)"       
#>  [46] "mALT(L)"           "LAL(L)"            "SPS(R)"           
#>  [49] "AL-DM1(L)"         "b'1(R)"            "PB"               
#>  [52] "ME(R)"             "AL-DC2(L)"         "FBl3"             
#>  [55] "PB(L5)"            "PRW"               "SIP(L)"           
#>  [58] "CRE(R)"            "VES(L)"            "AL-DM2(R)"        
#>  [61] "g2(R)"             "AL-VL2a(R)"        "PB(R7)"           
#>  [64] "AL-DL5(L)"         "VLNP(R)"           "SCL(L)"           
#>  [67] "mALT(R)"           "BU(R)"             "AL-DA2(L)"        
#>  [70] "NO"                "MB(L)"             "AL(L)"            
#>  [73] "MB(+ACA)(R)"       "AL-DC3(R)"         "AL-VP1l(R)"       
#>  [76] "PB(L1)"            "AL-DA4m(L)"        "FBl7"             
#>  [79] "ATL(R)"            "IPS(R)"            "AL-DA4l(R)"       
#>  [82] "AL-VC3l(R)"        "NO3(R)"            "EBr2r4"           
#>  [85] "NO2(R)"            "AL-VL2p(R)"        "AL-VM5d(R)"       
#>  [88] "PB(R5)"            "SMP(L)"            "gL(R)"            
#>  [91] "PB(R4)"            "AMMC"              "PENP"             
#>  [94] "OL(R)"             "AL-V(R)"           "AL-VP1d(R)"       
#>  [97] "CA(L)"             "LX(R)"             "PB(L9)"           
#> [100] "CAN(R)"            "g4(R)"             "AL-DP1m(L)"       
#> [103] "PB(R3)"            "GOR(R)"            "NO(L)"            
#> [106] "a1(R)"             "PB(R8)"            "EPA(R)"           
#> [109] "g5(R)"             "EBr3am"            "GOR(L)"           
#> [112] "b'L(R)"            "AL-VM7d(L)"        "AL-DM6(R)"        
#> [115] "AL-VP2(R)"         "AL-VP1m(R)"        "AL-VA6(R)"        
#> [118] "ATL(L)"            "AL-DL4(L)"         "FB-column3"       
#> [121] "a3(R)"             "SAD"               "EB"               
#> [124] "IB112"             "AL-VP4(R)"         "VES(R)"           
#> [127] "EBr6"              "AL-DM2(L)"         "PB(L4)"           
#> [130] "AL-VM7v(L)"        "FBl9"              "AL-DM5(R)"        
#> [133] "AL-VA3(R)"         "GF(R)"             "g3(R)"            
#> [136] "NO1(L)"            "PB(L6)"            "PVLP(R)"          
#> [139] "CRE(-ROB,-RUB)(R)" "FBl1"              "AL-DM4(R)"        
#> [142] "a'2(R)"            "FBl5"              "AL-DA4m(R)"       
#> [145] "AL-DP1l(R)"        "vACA(R)"           "PB(R6)"           
#> [148] "GA(R)"             "AB(R)"             "AL-DL2v(R)"       
#> [151] "AL-VC3m(R)"        "ICL(L)"            "PB(L8)"           
#> [154] "AL-DM6(L)"         "FBl8"              "AL(R)"            
#> [157] "AL-VA1v(R)"        "SNP(R)"            "AL-DA1(R)"        
#> [160] "AL-VC1(R)"         "SCL(R)"            "dACA(R)"          
#> [163] "AL-VL1(R)"         "VMNP"              "SPS(L)"           
#> [166] "AL-DM4(L)"         "AL-VM2(R)"         "aL(R)"            
#> [169] "b2(R)"             "AL-DC4(L)"         "AL-DA3(R)"        
#> [172] "WED(R)"            "PB(L7)"            "SMP(R)"           
#> [175] "b'2(R)"            "FBl6"              "AL-DL4(R)"        
#> [178] "AL-VA1d(R)"        "AL-VM3(R)"         "EBr1"             
#> [181] "AVLP(R)"           "IB"                "AL-DC2(R)"        
#> [184] "FB"                "AL-DM3(L)"         "SNP(L)"           
#> [187] "LO(R)"             "GNG"               "SLP(R)"           
#> [190] "CRE(-RUB)(L)"      "AL-VA4(R)"         "AL-VC2(R)"        
#> [193] "AL-VP5(R)"         "AOT(R)"            "a'L(R)"           
#> [196] "LAL(R)"            "PB(L2)"            "PB(R2)"           
#> [199] "a2(R)"             "AL-DC1(L)"         "NO(R)"            
#> [202] "CRE(L)"            "RUB(L)"            "AB(L)"            
#> [205] "a'1(R)"            "EBr3pw"            "SAD(-AMMC)"       
#> [208] "EPA(L)"            "AME(R)"            "AL-DL1(R)"        
#> [211] "AL-VA2(R)"         "AL-VA7l(R)"        "AL-VC4(R)"        
#> [214] "FBl4"              "AL-VA7m(R)"        "g1(R)"            
#> [217] "AL-DM1(R)"         "AL-DM3(R)"         "AL-VC5(R)"        
#> [220] "EBr5"              "EBr3d"             "AL-DA3(L)"        
#> [223] "AL-VA6(L)"         "PB(R9)"            "AL-DL2d(R)"       
#> [226] "AL-VA5(R)"         "AL-VM5v(R)"        "FLA(R)"           
#> [229] "INP"               "AL-D(R)"           "AL-VM1(R)"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "LAL(L)"            "mALT(R)"           "b2(R)"            
#>   [4] "ME(R)"             "FBl2"              "GNG"              
#>   [7] "LAL(R)"            "EBr3am"            "SPS(R)"           
#>  [10] "WED(R)"            "PB(L1)"            "a'1(R)"           
#>  [13] "AB(R)"             "a'3(R)"            "GF(R)"            
#>  [16] "g3(R)"             "SCL(R)"            "a'L(R)"           
#>  [19] "NO(R)"             "a'L(L)"            "PED(R)"           
#>  [22] "ROB(R)"            "MB(+ACA)(R)"       "INP"              
#>  [25] "NO2(L)"            "AOTU(R)"           "NO3(R)"           
#>  [28] "NO1(L)"            "VMNP"              "g4(R)"            
#>  [31] "FBl1"              "FBl3"              "CAN(R)"           
#>  [34] "PB"                "FBl4"              "FB"               
#>  [37] "PB(L7)"            "SCL(L)"            "CRE(R)"           
#>  [40] "PB(R8)"            "PB(R1)"            "CRE(L)"           
#>  [43] "PRW"               "FBl8"              "FBl5"             
#>  [46] "AVLP(R)"           "GOR(R)"            "EBr5"             
#>  [49] "PB(L8)"            "LX(R)"             "EBr3pw"           
#>  [52] "NO2(R)"            "PB(R4)"            "PB(L5)"           
#>  [55] "EBr2r4"            "EPA(L)"            "RUB(L)"           
#>  [58] "PB(L4)"            "IB"                "b'2(R)"           
#>  [61] "SMP(L)"            "g5(R)"             "SNP(L)"           
#>  [64] "LH(R)"             "vACA(R)"           "EB"               
#>  [67] "FBl9"              "OL(R)"             "SLP(R)"           
#>  [70] "SMP(R)"            "mALT(L)"           "a2(R)"            
#>  [73] "PB(L6)"            "CRE(-ROB,-RUB)(R)" "ATL(L)"           
#>  [76] "FBl6"              "NO(L)"             "PB(L3)"           
#>  [79] "hemibrain"         "dACA(R)"           "NO1(R)"           
#>  [82] "LO(R)"             "AME(R)"            "VLNP(R)"          
#>  [85] "g2(R)"             "EBr6"              "BU(R)"            
#>  [88] "AL(L)"             "PB(L9)"            "PB(R2)"           
#>  [91] "RUB(R)"            "b'L(R)"            "BU(L)"            
#>  [94] "FLA(R)"            "EBr3d"             "bL(L)"            
#>  [97] "PB(R5)"            "SIP(L)"            "SPS(L)"           
#> [100] "SNP(R)"            "EPA(R)"            "gL(R)"            
#> [103] "PVLP(R)"           "ATL(R)"            "b1(R)"            
#> [106] "AB(L)"             "GA(R)"             "PB(R7)"           
#> [109] "PB(R3)"            "a3(R)"             "MB(R)"            
#> [112] "b'1(R)"            "NO"                "AL(R)"            
#> [115] "LOP(R)"            "g1(R)"             "SIP(R)"           
#> [118] "IPS(R)"            "bL(R)"             "CX"               
#> [121] "CA(L)"             "LAL(-GA)(R)"       "GC"               
#> [124] "a'2(R)"            "gL(L)"             "aL(L)"            
#> [127] "b'L(L)"            "PB(R6)"            "PLP(R)"           
#> [130] "SAD(-AMMC)"        "CA(R)"             "AOT(R)"           
#> [133] "lACA(R)"           "PB(R9)"            "ICL(L)"           
#> [136] "VES(R)"            "LX(L)"             "VES(L)"           
#> [139] "POC"               "PENP"              "GOR(L)"           
#> [142] "EBr1"              "ICL(R)"            "MB(L)"            
#> [145] "NO3(L)"            "a1(R)"             "SAD"              
#> [148] "aL(R)"             "FBl7"              "PB(L2)"           
#> [151] "AMMC"             
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "mALT(R)"           "b2(R)"             "EPA(L)"           
#>   [4] "AL-VM1(R)"         "VES(R)"            "INP"              
#>   [7] "CRE(R)"            "AL-DC4(L)"         "AL-VP5(R)"        
#>  [10] "VLNP(R)"           "AL-DM4(L)"         "AL-VM7d(R)"       
#>  [13] "AL-VP1l(R)"        "VMNP"              "AL-DM5(L)"        
#>  [16] "a3(R)"             "a1(R)"             "ME(R)"            
#>  [19] "AL-DA2(L)"         "ROB(R)"            "AL-DM1(R)"        
#>  [22] "AL-DL1(R)"         "AB(L)"             "PB(R6)"           
#>  [25] "NO3(R)"            "AL(L)"             "FB-column3"       
#>  [28] "AL-DM2(L)"         "AL-VL2p(R)"        "LOP(R)"           
#>  [31] "dACA(R)"           "FLA(R)"            "FBl2"             
#>  [34] "GOR(R)"            "ATL(L)"            "PB"               
#>  [37] "FBl1"              "FBl3"              "AL-VC1(R)"        
#>  [40] "AL-DC4(R)"         "LAL(L)"            "NO1(L)"           
#>  [43] "LH(R)"             "NO(L)"             "AL-VM7d(L)"       
#>  [46] "SIP(R)"            "RUB(L)"            "PB(L4)"           
#>  [49] "aL(R)"             "BU(L)"             "AL-VA6(L)"        
#>  [52] "NO"                "AL-VM7v(R)"        "a'3(R)"           
#>  [55] "PB(L5)"            "PB(R3)"            "AL-VA6(R)"        
#>  [58] "AL-VA7m(R)"        "AL-VM4(R)"         "GF(R)"            
#>  [61] "PRW"               "CA(R)"             "SLP(R)"           
#>  [64] "AL-DL5(R)"         "b'L(R)"            "AL-DL5(L)"        
#>  [67] "g5(R)"             "FBl4"              "AL-VA1v(R)"       
#>  [70] "AL-DC1(L)"         "a'1(R)"            "AL-VA7l(R)"       
#>  [73] "AL-VM5v(R)"        "gL(L)"             "EBr3pw"           
#>  [76] "AL-VP3(R)"         "SNP(L)"            "FBl9"             
#>  [79] "aL(L)"             "GNG"               "CRE(L)"           
#>  [82] "ICL(R)"            "AL-VC3l(R)"        "MB(R)"            
#>  [85] "BU(R)"             "PB(R1)"            "AL-VM3(R)"        
#>  [88] "EPA(R)"            "AL-VM2(R)"         "AL-DA4m(L)"       
#>  [91] "CRE(-ROB,-RUB)(R)" "AOTU(R)"           "GA(R)"            
#>  [94] "AL-DA3(R)"         "b'1(R)"            "FBl6"             
#>  [97] "FBl7"              "AL-DA2(R)"         "SMP(R)"           
#> [100] "EBr3d"             "g4(R)"             "AL-DL4(L)"        
#> [103] "vACA(R)"           "AMMC"              "AL(R)"            
#> [106] "g1(R)"             "EBr1"              "PB(L2)"           
#> [109] "GC"                "FBl8"              "AL-DL2d(R)"       
#> [112] "AL-DM3(L)"         "ATL(R)"            "AL-VA4(R)"        
#> [115] "AL-VA5(R)"         "mALT(L)"           "b'2(R)"           
#> [118] "bL(R)"             "a'L(L)"            "EBr3am"           
#> [121] "AL-VP1d(R)"        "MB(L)"             "PB(R8)"           
#> [124] "PB(R9)"            "SAD(-AMMC)"        "b'L(L)"           
#> [127] "gL(R)"             "PB(L9)"            "MB(+ACA)(R)"      
#> [130] "PB(R2)"            "SCL(L)"            "GOR(L)"           
#> [133] "PB(L7)"            "AL-DC1(R)"         "POC"              
#> [136] "hemibrain"         "AL-DM6(L)"         "AL-DC2(R)"        
#> [139] "AL-VC5(R)"         "PENP"              "AL-VM7v(L)"       
#> [142] "LO(R)"             "CAN(R)"            "NO3(L)"           
#> [145] "AL-DC3(R)"         "AL-VL1(R)"         "b1(R)"            
#> [148] "AL-VC3m(R)"        "a'L(R)"            "AL-VM5d(R)"       
#> [151] "AL-DA3(L)"         "AL-DC2(L)"         "FBl5"             
#> [154] "AL-DP1l(R)"        "PB(R5)"            "CA(L)"            
#> [157] "SPS(L)"            "LX(R)"             "a'2(R)"           
#> [160] "AL-VP4(R)"         "SAD"               "EBr5"             
#> [163] "NO2(R)"            "LAL(-GA)(R)"       "EBr6"             
#> [166] "EBr2r4"            "RUB(R)"            "PED(R)"           
#> [169] "AL-DM5(R)"         "AL-D(R)"           "AL-VC4(R)"        
#> [172] "IB"                "g3(R)"             "NO(R)"            
#> [175] "AL-DM2(R)"         "AL-DM3(R)"         "PB(L3)"           
#> [178] "CX"                "AL-VP2(R)"         "SIP(L)"           
#> [181] "SCL(R)"            "IPS(R)"            "AL-DA1(R)"        
#> [184] "AL-VA3(R)"         "PB(L1)"            "CRE(-RUB)(L)"     
#> [187] "PB(R4)"            "AL-VL2a(R)"        "AOT(R)"           
#> [190] "g2(R)"             "bL(L)"             "LX(L)"            
#> [193] "AL-DL4(R)"         "SMP(L)"            "a2(R)"            
#> [196] "AL-D(L)"           "AL-DP1m(L)"        "LAL(R)"           
#> [199] "AL-DM4(R)"         "VES(L)"            "AL-DA4l(R)"       
#> [202] "FB"                "ICL(L)"            "AB(R)"            
#> [205] "OL(R)"             "PLP(R)"            "PVLP(R)"          
#> [208] "AL-VP1m(R)"        "NO2(L)"            "WED(R)"           
#> [211] "AL-DA4m(R)"        "AL-DM6(R)"         "PB(R7)"           
#> [214] "EB"                "AL-DM1(L)"         "AL-DL3(R)"        
#> [217] "AL-VC2(R)"         "AL-V(R)"           "SPS(R)"           
#> [220] "SNP(R)"            "AL-DL2v(R)"        "AL-VA2(R)"        
#> [223] "lACA(R)"           "NO1(R)"            "AVLP(R)"          
#> [226] "AL-VA1d(R)"        "PB(L8)"            "PB(L6)"           
#> [229] "AME(R)"            "AL-DP1m(R)"       
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "GOR(L)"            "LO(R)"             "PED(R)"           
#>   [4] "INP"               "AL-VA4(R)"         "AL-DL5(R)"        
#>   [7] "AL-D(L)"           "AL-DM6(L)"         "AL-DM4(R)"        
#>  [10] "LAL(L)"            "LAL(-GA)(R)"       "ME(R)"            
#>  [13] "CRE(L)"            "AL-VP1d(R)"        "SMP(R)"           
#>  [16] "a1(R)"             "AL-DC4(L)"         "ATL(R)"           
#>  [19] "aL(L)"             "AL(R)"             "PENP"             
#>  [22] "AL(L)"             "ICL(L)"            "AL-VL2a(R)"       
#>  [25] "PB(R5)"            "AL-DA4m(L)"        "AME(R)"           
#>  [28] "vACA(R)"           "PB(L9)"            "VES(L)"           
#>  [31] "AL-VA5(R)"         "AL-VM4(R)"         "AMMC"             
#>  [34] "PB(R9)"            "AL-DP1l(R)"        "AOTU(R)"          
#>  [37] "a'2(R)"            "AL-DP1m(L)"        "bL(R)"            
#>  [40] "EBr3am"            "FB-column3"        "AL-VM7v(R)"       
#>  [43] "AL-DL4(L)"         "AL-DL5(L)"         "g5(R)"            
#>  [46] "FBl8"              "AL-DA4m(R)"        "IB"               
#>  [49] "AL-VM5d(R)"        "PB(R7)"            "SCL(R)"           
#>  [52] "BU(L)"             "FBl7"              "AL-VC4(R)"        
#>  [55] "AOT(R)"            "SAD"               "EB"               
#>  [58] "g4(R)"             "PB(L4)"            "AL-V(R)"          
#>  [61] "PB(L1)"            "EPA(R)"            "AL-DA4l(R)"       
#>  [64] "AL-DC4(R)"         "SIP(L)"            "AL-VA7m(R)"       
#>  [67] "AL-VC3l(R)"        "PB(L3)"            "GA(R)"            
#>  [70] "PRW"               "PB(R3)"            "RUB(L)"           
#>  [73] "AL-DM6(R)"         "AL-VM7d(R)"        "LX(R)"            
#>  [76] "NO(R)"             "bL(L)"             "MB(+ACA)(R)"      
#>  [79] "EPA(L)"            "NO1(R)"            "PB(R1)"           
#>  [82] "FBl2"              "AL-VA3(R)"         "AL-VP5(R)"        
#>  [85] "AL-VP1m(R)"        "GC"                "EBr3pw"           
#>  [88] "AL-DM2(L)"         "AL-DC2(R)"         "PB(R6)"           
#>  [91] "AL-VA6(L)"         "ROB(R)"            "a3(R)"            
#>  [94] "OL(R)"             "AL-VA7l(R)"        "PB(R8)"           
#>  [97] "AL-VM7d(L)"        "NO"                "LH(R)"            
#> [100] "aL(R)"             "FBl1"              "PB(R2)"           
#> [103] "AL-DL3(R)"         "CA(L)"             "EBr1"             
#> [106] "EBr3d"             "b1(R)"             "CRE(R)"           
#> [109] "MB(R)"             "AL-DC2(L)"         "VES(R)"           
#> [112] "GOR(R)"            "AL-D(R)"           "PB(L5)"           
#> [115] "lACA(R)"           "SNP(L)"            "CRE(-ROB,-RUB)(R)"
#> [118] "b'2(R)"            "SPS(R)"            "AL-DL2d(R)"       
#> [121] "mALT(L)"           "AL-DM3(L)"         "a'L(L)"           
#> [124] "POC"               "SCL(L)"            "AL-DM5(R)"        
#> [127] "AL-VC1(R)"         "AL-VM2(R)"         "AL-VM3(R)"        
#> [130] "hemibrain"         "mALT(R)"           "AL-DL2v(R)"       
#> [133] "AL-VP2(R)"         "a'3(R)"            "ATL(L)"           
#> [136] "a'1(R)"            "NO2(R)"            "a'L(R)"           
#> [139] "gL(L)"             "AL-DA1(R)"         "EBr6"             
#> [142] "SNP(R)"            "CRE(-RUB)(L)"      "AL-DP1m(R)"       
#> [145] "SPS(L)"            "AL-DC1(L)"         "AL-VM7v(L)"       
#> [148] "MB(L)"             "GNG"               "NO3(R)"           
#> [151] "EBr2r4"            "AL-DA2(L)"         "AL-VA2(R)"        
#> [154] "g1(R)"             "AL-VA1v(R)"        "AL-VM1(R)"        
#> [157] "g3(R)"             "g2(R)"             "FBl9"             
#> [160] "AL-DM2(R)"         "EBr5"              "AL-DM4(L)"        
#> [163] "PB(L8)"            "ICL(R)"            "AL-DM1(R)"        
#> [166] "a2(R)"             "b2(R)"             "LX(L)"            
#> [169] "AL-DA3(R)"         "AL-VC3m(R)"        "FB"               
#> [172] "dACA(R)"           "PB"                "PLP(R)"           
#> [175] "PB(L7)"            "AL-VA6(R)"         "b'L(R)"           
#> [178] "AL-DM1(L)"         "PB(L6)"            "LAL(R)"           
#> [181] "AL-VP4(R)"         "BU(R)"             "AL-DL4(R)"        
#> [184] "AL-VM5v(R)"        "CX"                "AL-VL1(R)"        
#> [187] "AL-VP1l(R)"        "AB(L)"             "LOP(R)"           
#> [190] "SIP(R)"            "b'L(L)"            "FBl4"             
#> [193] "FLA(R)"            "VMNP"              "AL-DA3(L)"        
#> [196] "SAD(-AMMC)"        "FBl6"              "SLP(R)"           
#> [199] "AL-DA2(R)"         "AL-VC5(R)"         "AL-DM5(L)"        
#> [202] "FBl5"              "AL-DC1(R)"         "NO2(L)"           
#> [205] "SMP(L)"            "AL-VP3(R)"         "PB(R4)"           
#> [208] "GF(R)"             "AVLP(R)"           "AL-VA1d(R)"       
#> [211] "AL-VL2p(R)"        "RUB(R)"            "b'1(R)"           
#> [214] "AL-VC2(R)"         "NO1(L)"            "CA(R)"            
#> [217] "IPS(R)"            "NO(L)"             "AL-DM3(R)"        
#> [220] "gL(R)"             "FBl3"              "AB(R)"            
#> [223] "CAN(R)"            "AL-DL1(R)"         "NO3(L)"           
#> [226] "VLNP(R)"           "PB(L2)"            "WED(R)"           
#> [229] "PVLP(R)"           "AL-DC3(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                          
#> [1,] "POST" "/api/npexplorer/roisinneuron"                
#> [2,] "GET"  "/api/npexplorer/celltype/:dataset/:type"     
#> [3,] "POST" "/api/v:ver/raw/cypher/transaction"           
#> [4,] "POST" "/api/v:ver/raw/cypher/transaction/:id/cypher"
#> [5,] "POST" "/api/v:ver/raw/keyvalue/key/:instance/:key"  
#> [6,] "POST" "/logout"                                     
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIsInNeuron-fm"
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCellType-fm"    
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [6,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                    
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }