Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AB(R)"             "AL-VM5d(R)"        "AL-DM6(L)"        
#>   [4] "AL-DC1(R)"         "ATL(L)"            "CRE(R)"           
#>   [7] "AL-DC4(L)"         "PB(L7)"            "FBl4"             
#>  [10] "FBl5"              "AL-VP4(R)"         "AL-DA4l(R)"       
#>  [13] "PRW"               "PB"                "SIP(R)"           
#>  [16] "OL(R)"             "AL-VC3l(R)"        "SAD(-AMMC)"       
#>  [19] "IB112"             "FBl2"              "AL-DC2(R)"        
#>  [22] "AL-V(R)"           "g5(R)"             "CA(R)"            
#>  [25] "VES(L)"            "a'L(R)"            "AL-DM5(L)"        
#>  [28] "AL-DM2(R)"         "AL-DM6(R)"         "AL-VA7l(R)"       
#>  [31] "AL-VM7v(R)"        "AL-VP1l(R)"        "a2(R)"            
#>  [34] "lACA(R)"           "aL(R)"             "MB(L)"            
#>  [37] "EBr6"              "PB(R2)"            "AL-VC4(R)"        
#>  [40] "NO3(R)"            "AL-DA2(L)"         "AL-DA4m(L)"       
#>  [43] "AL-VM7d(R)"        "BU(L)"             "gL(R)"            
#>  [46] "PVLP(R)"           "CAN(R)"            "AL-DA4m(R)"       
#>  [49] "a3(R)"             "AL-DP1m(R)"        "AL-VA5(R)"        
#>  [52] "AL-VL1(R)"         "AL-DM2(L)"         "AL-DC2(L)"        
#>  [55] "AL-DA3(R)"         "PB(R5)"            "PLP(R)"           
#>  [58] "MB(R)"             "AL-VA4(R)"         "GA(R)"            
#>  [61] "SIP(L)"            "AL-VM7v(L)"        "ATL(R)"           
#>  [64] "GOR(R)"            "b'L(L)"            "AL-DA2(R)"        
#>  [67] "AL-VP1m(R)"        "PED(R)"            "SCL(R)"           
#>  [70] "AME(R)"            "AL-VM2(R)"         "gL(L)"            
#>  [73] "AL(L)"             "FB-column3"        "AL-VA6(L)"        
#>  [76] "PB(L2)"            "FBl6"              "AL-DM1(R)"        
#>  [79] "AL-VP2(R)"         "g3(R)"             "EBr2r4"           
#>  [82] "BU(R)"             "SLP(R)"            "AL-DP1l(R)"       
#>  [85] "AL-VA2(R)"         "AL-VC1(R)"         "PB(R7)"           
#>  [88] "AL-VL2p(R)"        "AL-VM1(R)"         "VLNP(R)"          
#>  [91] "LAL(L)"            "PB(L6)"            "NO3(L)"           
#>  [94] "AL-DL5(R)"         "AL-VA3(R)"         "AL-DM1(L)"        
#>  [97] "PB(L9)"            "AL-DL1(R)"         "AL-DL4(R)"        
#> [100] "AL-VA7m(R)"        "b'L(R)"            "ICL(R)"           
#> [103] "AB(L)"             "NO2(L)"            "EBr5"             
#> [106] "NO1(R)"            "g4(R)"             "NO2(R)"           
#> [109] "FBl9"              "EBr3am"            "PB(R3)"           
#> [112] "a'L(L)"            "AL-VP1d(R)"        "WED(R)"           
#> [115] "PB(L4)"            "AL-VP3(R)"         "NO1(L)"           
#> [118] "FB"                "EBr3pw"            "LO(R)"            
#> [121] "GF(R)"             "dACA(R)"           "PB(L3)"           
#> [124] "AL-VC5(R)"         "SMP(R)"            "IB"               
#> [127] "NO(R)"             "AL-DL2v(R)"        "b2(R)"            
#> [130] "g2(R)"             "PB(R8)"            "AL-DL4(L)"        
#> [133] "VES(R)"            "LAL(-GA)(R)"       "SPS(L)"           
#> [136] "CRE(L)"            "RUB(L)"            "ICL(L)"           
#> [139] "MB(+ACA)(R)"       "aL(L)"             "AL-VP5(R)"        
#> [142] "b'1(R)"            "LOP(R)"            "PB(L5)"           
#> [145] "PB(R6)"            "b'2(R)"            "AOTU(R)"          
#> [148] "GC"                "GNG"               "SMP(L)"           
#> [151] "AL-DA3(L)"         "IPS(R)"            "AL-VA1v(R)"       
#> [154] "CRE(-ROB,-RUB)(R)" "CX"                "EBr1"             
#> [157] "EB"                "LX(L)"             "AMMC"             
#> [160] "mALT(R)"           "NO"                "AL-VA1d(R)"       
#> [163] "GOR(L)"            "LAL(R)"            "AL-DC3(R)"        
#> [166] "ME(R)"             "AL-DM5(R)"         "hemibrain"        
#> [169] "EBr3d"             "AL-DP1m(L)"        "CA(L)"            
#> [172] "mALT(L)"           "PB(L1)"            "FLA(R)"           
#> [175] "AL-D(L)"           "AL-DM3(L)"         "AL(R)"            
#> [178] "AL-VC2(R)"         "AOT(R)"            "bL(R)"            
#> [181] "AL-VC3m(R)"        "vACA(R)"           "a'1(R)"           
#> [184] "PB(R4)"            "AL-DL3(R)"         "AL-DM3(R)"        
#> [187] "AL-DM4(R)"         "AL-VA6(R)"         "AL-VM4(R)"        
#> [190] "SNP(R)"            "AL-DL2d(R)"        "AL-VM5v(R)"       
#> [193] "PENP"              "g1(R)"             "AL-DM4(L)"        
#> [196] "LH(R)"             "AL-D(R)"           "EPA(R)"           
#> [199] "AL-DC1(L)"         "EPA(L)"            "INP"              
#> [202] "SCL(L)"            "a1(R)"             "CRE(-RUB)(L)"     
#> [205] "FBl3"              "SAD"               "AL-DL5(L)"        
#> [208] "PB(R9)"            "AVLP(R)"           "PB(L8)"           
#> [211] "LX(R)"             "FBl1"              "a'3(R)"           
#> [214] "SPS(R)"            "AL-DC4(R)"         "AL-VM3(R)"        
#> [217] "ROB(R)"            "PB(R1)"            "FBl7"             
#> [220] "NO(L)"             "AL-DA1(R)"         "FBl8"             
#> [223] "POC"               "bL(L)"             "RUB(R)"           
#> [226] "AL-VM7d(L)"        "b1(R)"             "AL-VL2a(R)"       
#> [229] "SNP(L)"            "VMNP"              "a'2(R)"           
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PB(R4)"            "a'L(L)"            "vACA(R)"          
#>   [4] "hemibrain"         "SIP(L)"            "a'1(R)"           
#>   [7] "aL(L)"             "LO(R)"             "CAN(R)"           
#>  [10] "LOP(R)"            "EBr3pw"            "VES(R)"           
#>  [13] "IPS(R)"            "GA(R)"             "mALT(R)"          
#>  [16] "WED(R)"            "SCL(R)"            "gL(R)"            
#>  [19] "PB(L9)"            "FBl6"              "MB(R)"            
#>  [22] "SLP(R)"            "PB(L5)"            "mALT(L)"          
#>  [25] "aL(R)"             "PB(L1)"            "SAD(-AMMC)"       
#>  [28] "RUB(L)"            "lACA(R)"           "GOR(R)"           
#>  [31] "PB(L4)"            "FBl7"              "OL(R)"            
#>  [34] "SMP(R)"            "g1(R)"             "MB(L)"            
#>  [37] "ROB(R)"            "FBl4"              "SMP(L)"           
#>  [40] "LX(R)"             "PB(R2)"            "bL(R)"            
#>  [43] "CA(L)"             "NO2(L)"            "g3(R)"            
#>  [46] "a'L(R)"            "FBl3"              "GNG"              
#>  [49] "PED(R)"            "CRE(-ROB,-RUB)(R)" "a2(R)"            
#>  [52] "NO1(L)"            "POC"               "AB(R)"            
#>  [55] "INP"               "AMMC"              "PB(L8)"           
#>  [58] "PLP(R)"            "FBl8"              "LH(R)"            
#>  [61] "b'2(R)"            "SNP(L)"            "FBl1"             
#>  [64] "b'1(R)"            "PB(R7)"            "GOR(L)"           
#>  [67] "BU(R)"             "NO(R)"             "PB(R8)"           
#>  [70] "AL(R)"             "CX"                "MB(+ACA)(R)"      
#>  [73] "PENP"              "SAD"               "NO"               
#>  [76] "PVLP(R)"           "g4(R)"             "EPA(R)"           
#>  [79] "EBr3am"            "LAL(-GA)(R)"       "FLA(R)"           
#>  [82] "PB"                "SNP(R)"            "SPS(R)"           
#>  [85] "b1(R)"             "VLNP(R)"           "EBr2r4"           
#>  [88] "bL(L)"             "PB(L7)"            "NO2(R)"           
#>  [91] "PRW"               "g2(R)"             "BU(L)"            
#>  [94] "EBr3d"             "VES(L)"            "a1(R)"            
#>  [97] "EB"                "AL(L)"             "RUB(R)"           
#> [100] "AOT(R)"            "PB(L3)"            "EBr6"             
#> [103] "CA(R)"             "NO(L)"             "PB(R9)"           
#> [106] "ATL(R)"            "AOTU(R)"           "GC"               
#> [109] "FBl2"              "LAL(R)"            "PB(R5)"           
#> [112] "PB(R6)"            "CRE(L)"            "AVLP(R)"          
#> [115] "AME(R)"            "gL(L)"             "EPA(L)"           
#> [118] "SIP(R)"            "b'L(L)"            "b'L(R)"           
#> [121] "EBr5"              "NO1(R)"            "ME(R)"            
#> [124] "NO3(R)"            "SPS(L)"            "a'2(R)"           
#> [127] "PB(L2)"            "EBr1"              "NO3(L)"           
#> [130] "LX(L)"             "VMNP"              "IB"               
#> [133] "a'3(R)"            "GF(R)"             "b2(R)"            
#> [136] "FB"                "dACA(R)"           "a3(R)"            
#> [139] "LAL(L)"            "CRE(R)"            "PB(R1)"           
#> [142] "ICL(R)"            "FBl9"              "PB(R3)"           
#> [145] "AB(L)"             "ATL(L)"            "PB(L6)"           
#> [148] "ICL(L)"            "g5(R)"             "FBl5"             
#> [151] "SCL(L)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "LAL(L)"            "aL(R)"             "FLA(R)"           
#>   [4] "BU(R)"             "AME(R)"            "AL-VA1d(R)"       
#>   [7] "EBr3am"            "AL-DP1m(L)"        "SAD(-AMMC)"       
#>  [10] "AL-D(R)"           "EBr2r4"            "AL-DM3(L)"        
#>  [13] "FBl6"              "AB(L)"             "AL-DA4m(L)"       
#>  [16] "IPS(R)"            "AL-DM6(R)"         "AL-VP3(R)"        
#>  [19] "AL-VM7v(R)"        "AL-DC1(L)"         "PB(L9)"           
#>  [22] "FBl4"              "AL-DL2d(R)"        "AL-DM2(R)"        
#>  [25] "CRE(R)"            "AL-D(L)"           "a'1(R)"           
#>  [28] "EPA(L)"            "AL-VM5v(R)"        "AL-DA3(R)"        
#>  [31] "AL-DM4(L)"         "SAD"               "b2(R)"            
#>  [34] "BU(L)"             "AL-DM5(R)"         "AL-VM2(R)"        
#>  [37] "GA(R)"             "MB(+ACA)(R)"       "AL-VP2(R)"        
#>  [40] "AL-DM2(L)"         "AL-VA6(L)"         "AL-DM4(R)"        
#>  [43] "AL-VC3l(R)"        "AMMC"              "ROB(R)"           
#>  [46] "vACA(R)"           "NO3(R)"            "CRE(L)"           
#>  [49] "AL-DL3(R)"         "AL-VC1(R)"         "AL-VC3m(R)"       
#>  [52] "FBl1"              "AL-VA4(R)"         "PB"               
#>  [55] "FB-column3"        "AL-DL1(R)"         "VLNP(R)"          
#>  [58] "PB(L5)"            "EBr6"              "a'2(R)"           
#>  [61] "LH(R)"             "AL-VA5(R)"         "AL-VC2(R)"        
#>  [64] "LAL(-GA)(R)"       "PB(L8)"            "EB"               
#>  [67] "AL-DA4m(R)"        "AL-VC4(R)"         "EBr3d"            
#>  [70] "NO2(R)"            "CA(L)"             "b'L(R)"           
#>  [73] "AL-DL4(L)"         "NO(L)"             "AL-DL2v(R)"       
#>  [76] "AL-VA6(R)"         "PB(L3)"            "IB"               
#>  [79] "ME(R)"             "NO"                "AL-DC2(R)"        
#>  [82] "AL-DP1m(R)"        "AL-DM6(L)"         "SLP(R)"           
#>  [85] "AL(L)"             "CRE(-RUB)(L)"      "AL-DA2(R)"        
#>  [88] "CX"                "NO2(L)"            "g4(R)"            
#>  [91] "AL-DC4(L)"         "g5(R)"             "FBl5"             
#>  [94] "AL-VL2a(R)"        "hemibrain"         "dACA(R)"          
#>  [97] "LX(R)"             "AL-VA3(R)"         "RUB(R)"           
#> [100] "CRE(-ROB,-RUB)(R)" "PENP"              "PB(R6)"           
#> [103] "SNP(R)"            "PB(L6)"            "AL-VA7l(R)"       
#> [106] "MB(R)"             "POC"               "mALT(L)"          
#> [109] "NO(R)"             "OL(R)"             "SNP(L)"           
#> [112] "PB(R3)"            "AL-VA7m(R)"        "AL-DM1(L)"        
#> [115] "FBl7"              "AL-VM7d(R)"        "lACA(R)"          
#> [118] "AL-DA2(L)"         "PB(L7)"            "FBl9"             
#> [121] "INP"               "FBl8"              "AB(R)"            
#> [124] "CAN(R)"            "AL-DM3(R)"         "AL-DP1l(R)"       
#> [127] "FB"                "SMP(R)"            "g1(R)"            
#> [130] "AL-DC1(R)"         "AL-VA1v(R)"        "AL-VL1(R)"        
#> [133] "AL-VP5(R)"         "WED(R)"            "EPA(R)"           
#> [136] "a1(R)"             "a'3(R)"            "ATL(R)"           
#> [139] "AL-VL2p(R)"        "AL-VP1m(R)"        "LOP(R)"           
#> [142] "EBr3pw"            "CA(R)"             "SPS(R)"           
#> [145] "mALT(R)"           "MB(L)"             "EBr5"             
#> [148] "AL-DA3(L)"         "RUB(L)"            "AL-DL5(R)"        
#> [151] "g3(R)"             "EBr1"              "PB(L1)"           
#> [154] "GOR(R)"            "LAL(R)"            "NO3(L)"           
#> [157] "AL-VM7d(L)"        "PB(R1)"            "FBl3"             
#> [160] "AL-V(R)"           "a2(R)"             "g2(R)"            
#> [163] "a3(R)"             "SCL(L)"            "SCL(R)"           
#> [166] "a'L(R)"            "PB(R2)"            "AL-VA2(R)"        
#> [169] "AOT(R)"            "gL(L)"             "bL(R)"            
#> [172] "PLP(R)"            "AL-DM5(L)"         "FBl2"             
#> [175] "b'L(L)"            "AL-DA4l(R)"        "GC"               
#> [178] "VES(R)"            "b'2(R)"            "AL-DC2(L)"        
#> [181] "PRW"               "AL-VP4(R)"         "ATL(L)"           
#> [184] "b'1(R)"            "SMP(L)"            "SPS(L)"           
#> [187] "GNG"               "AL-DL4(R)"         "AL-DM1(R)"        
#> [190] "AOTU(R)"           "LX(L)"             "AL-DA1(R)"        
#> [193] "AL-DC3(R)"         "AL-VP1l(R)"        "VMNP"             
#> [196] "PB(R9)"            "PB(R4)"            "a'L(L)"           
#> [199] "AL-DC4(R)"         "PB(R8)"            "PB(R5)"           
#> [202] "SIP(L)"            "SIP(R)"            "gL(R)"            
#> [205] "ICL(L)"            "aL(L)"             "NO1(R)"           
#> [208] "ICL(R)"            "PVLP(R)"           "GF(R)"            
#> [211] "PB(R7)"            "PED(R)"            "GOR(L)"           
#> [214] "NO1(L)"            "AL-VM4(R)"         "AL-DL5(L)"        
#> [217] "bL(L)"             "PB(L4)"            "AL-VM3(R)"        
#> [220] "PB(L2)"            "AVLP(R)"           "VES(L)"           
#> [223] "AL(R)"             "AL-VM1(R)"         "AL-VP1d(R)"       
#> [226] "AL-VM7v(L)"        "LO(R)"             "AL-VM5d(R)"       
#> [229] "b1(R)"             "AL-VC5(R)"        
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "BU(L)"             "bL(R)"             "AL-DL3(R)"        
#>   [4] "AL-VA7l(R)"        "hemibrain"         "AL-DC4(R)"        
#>   [7] "AL-D(R)"           "g4(R)"             "NO(R)"            
#>  [10] "GNG"               "AL-DC2(R)"         "AL-V(R)"          
#>  [13] "NO2(L)"            "PB(R1)"            "LH(R)"            
#>  [16] "CAN(R)"            "EPA(L)"            "FBl4"             
#>  [19] "FBl7"              "LAL(L)"            "AL-DL5(L)"        
#>  [22] "CRE(L)"            "SNP(L)"            "SLP(R)"           
#>  [25] "AL-DL2d(R)"        "NO2(R)"            "AOT(R)"           
#>  [28] "SMP(L)"            "AL-DL4(L)"         "SMP(R)"           
#>  [31] "bL(L)"             "AL-DA1(R)"         "AL-DM4(R)"        
#>  [34] "ATL(L)"            "LX(R)"             "AL-DM5(R)"        
#>  [37] "AL-VC3l(R)"        "AL-VP5(R)"         "VLNP(R)"          
#>  [40] "GA(R)"             "PB(L4)"            "EBr3am"           
#>  [43] "g3(R)"             "b'L(R)"            "EBr6"             
#>  [46] "AL-DP1m(R)"        "PB(R8)"            "AL-VP1m(R)"       
#>  [49] "IB"                "AL-DM2(R)"         "PB(R6)"           
#>  [52] "AL-DM5(L)"         "NO(L)"             "AL-VC3m(R)"       
#>  [55] "AL-VM4(R)"         "MB(L)"             "AL-DC4(L)"        
#>  [58] "AL-DM3(L)"         "PB(L6)"            "a'1(R)"           
#>  [61] "AL-DC1(R)"         "PB(L5)"            "SIP(L)"           
#>  [64] "a'3(R)"            "AL-DC2(L)"         "AL-DL2v(R)"       
#>  [67] "AL-VP1l(R)"        "lACA(R)"           "CRE(-RUB)(L)"     
#>  [70] "FBl1"              "AB(R)"             "ME(R)"            
#>  [73] "LO(R)"             "AL-DL5(R)"         "AL-VM3(R)"        
#>  [76] "PB(R7)"            "SPS(R)"            "a'L(L)"           
#>  [79] "AL-DA3(R)"         "AL-VC1(R)"         "AL-VP2(R)"        
#>  [82] "AB(L)"             "SPS(L)"            "AL-DM4(L)"        
#>  [85] "AL-VL2a(R)"        "b2(R)"             "PB(L7)"           
#>  [88] "AL-VC2(R)"         "a1(R)"             "NO3(L)"           
#>  [91] "g5(R)"             "FBl6"              "AL-VL2p(R)"       
#>  [94] "GC"                "FBl8"              "AL-VL1(R)"        
#>  [97] "AL-VM5v(R)"        "AL-VP4(R)"         "vACA(R)"          
#> [100] "AL-VP1d(R)"        "PB(L1)"            "AL-DC1(L)"        
#> [103] "AL-VA1v(R)"        "AL-VC4(R)"         "PB(L9)"           
#> [106] "PVLP(R)"           "ATL(R)"            "a2(R)"            
#> [109] "NO"                "AL-DA4l(R)"        "FLA(R)"           
#> [112] "AL(L)"             "MB(+ACA)(R)"       "AL-DL1(R)"        
#> [115] "PB(L3)"            "PB(L8)"            "b'L(L)"           
#> [118] "AL-DM1(R)"         "AL-VM7v(R)"        "GOR(L)"           
#> [121] "ICL(L)"            "AL-VA1d(R)"        "EBr3pw"           
#> [124] "PENP"              "VMNP"              "SIP(R)"           
#> [127] "MB(R)"             "PLP(R)"            "RUB(L)"           
#> [130] "AL-DM3(R)"         "SNP(R)"            "SCL(R)"           
#> [133] "EB"                "AL-DA4m(L)"        "PB(R5)"           
#> [136] "mALT(L)"           "CA(L)"             "AMMC"             
#> [139] "WED(R)"            "EBr3d"             "AL-DA3(L)"        
#> [142] "AL-DM6(L)"         "AL-DP1l(R)"        "AL-VM5d(R)"       
#> [145] "RUB(R)"            "CRE(R)"            "AL-VA6(L)"        
#> [148] "AL-DL4(R)"         "AL-VA3(R)"         "SCL(L)"           
#> [151] "AL-D(L)"           "AL-VM7d(R)"        "SAD"              
#> [154] "dACA(R)"           "AVLP(R)"           "LX(L)"            
#> [157] "VES(L)"            "PB(L2)"            "PRW"              
#> [160] "PB(R4)"            "AL-VM2(R)"         "gL(L)"            
#> [163] "AL-VM7v(L)"        "LAL(R)"            "AL-VP3(R)"        
#> [166] "BU(R)"             "AL-DP1m(L)"        "b'1(R)"           
#> [169] "OL(R)"             "AL(R)"             "AL-DA4m(R)"       
#> [172] "LAL(-GA)(R)"       "g2(R)"             "PB"               
#> [175] "AL-DM2(L)"         "a3(R)"             "AL-DM1(L)"        
#> [178] "FBl9"              "GF(R)"             "FBl3"             
#> [181] "b1(R)"             "IPS(R)"            "AL-VA2(R)"        
#> [184] "AL-VA6(R)"         "SAD(-AMMC)"        "CA(R)"            
#> [187] "AME(R)"            "AL-DM6(R)"         "AL-VM7d(L)"       
#> [190] "ROB(R)"            "FB-column3"        "AL-DA2(R)"        
#> [193] "LOP(R)"            "g1(R)"             "NO1(L)"           
#> [196] "gL(R)"             "FBl5"              "PB(R2)"           
#> [199] "FB"                "FBl2"              "AL-DC3(R)"        
#> [202] "aL(R)"             "a'2(R)"            "a'L(R)"           
#> [205] "ICL(R)"            "GOR(R)"            "AL-VA5(R)"        
#> [208] "b'2(R)"            "AOTU(R)"           "mALT(R)"          
#> [211] "EBr1"              "POC"               "CX"               
#> [214] "EPA(R)"            "INP"               "AL-VA4(R)"        
#> [217] "CRE(-ROB,-RUB)(R)" "NO3(R)"            "NO1(R)"           
#> [220] "PB(R3)"            "aL(L)"             "AL-VA7m(R)"       
#> [223] "AL-VC5(R)"         "AL-VM1(R)"         "PED(R)"           
#> [226] "EBr2r4"            "VES(R)"            "EBr5"             
#> [229] "AL-DA2(L)"         "PB(R9)"           
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                      
#> [1,] "GET"  "/api/v:ver/npexplorer/commonconnectivity"
#> [2,] "GET"  "/api/npexplorer/distribution"            
#> [3,] "POST" "/api/v:ver/npexplorer/completeness"      
#> [4,] "GET"  "/api"                                    
#> [5,] "GET"  "/api/dbmeta/datasets"                    
#> [6,] "GET"  "/api/v:ver/dbmeta/instances"             
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [4,] "github.com/labstack/echo.glob..func1"                                                      
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getDatasets-fm"            
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }