Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0`
#> $`hemibrain:v1.0`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0`$ROIs
#>   [1] "MB(R)"             "PB(L2)"            "GF(R)"            
#>   [4] "GOR(L)"            "mALT(R)"           "PLP(R)"           
#>   [7] "a'1(R)"            "b1(R)"             "OL(R)"            
#>  [10] "SPS(R)"            "WED(R)"            "MB(L)"            
#>  [13] "BU(L)"             "FLA(R)"            "VES(L)"           
#>  [16] "PB(L3)"            "gL(L)"             "BU(R)"            
#>  [19] "AVLP(R)"           "CX"                "g4(R)"            
#>  [22] "SAD(-AMMC)"        "ICL(R)"            "b'L(L)"           
#>  [25] "AB(L)"             "AL(L)"             "ATL(R)"           
#>  [28] "LAL(R)"            "RUB(L)"            "g1(R)"            
#>  [31] "NO1(L)"            "dACA(R)"           "PVLP(R)"          
#>  [34] "FBl8"              "aL(L)"             "INP"              
#>  [37] "a1(R)"             "EBr1"              "a'L(R)"           
#>  [40] "CA(R)"             "a'2(R)"            "PB(L9)"           
#>  [43] "g5(R)"             "PB(R6)"            "hemibrain"        
#>  [46] "NO3(R)"            "EBr3d"             "CAN(R)"           
#>  [49] "NO(L)"             "b'1(R)"            "lACA(R)"          
#>  [52] "PB(R7)"            "LAL(-GA)(R)"       "PB(R8)"           
#>  [55] "g3(R)"             "EB"                "FBl1"             
#>  [58] "AOT(R)"            "mALT(L)"           "ATL(L)"           
#>  [61] "PENP"              "b'2(R)"            "AOTU(R)"          
#>  [64] "LH(R)"             "a3(R)"             "a'L(L)"           
#>  [67] "SLP(R)"            "SIP(L)"            "PB(L1)"           
#>  [70] "SIP(R)"            "PB(R1)"            "PB(R3)"           
#>  [73] "NO2(R)"            "PB(R4)"            "b'L(R)"           
#>  [76] "PB"                "PB(L6)"            "gL(R)"            
#>  [79] "ICL(L)"            "SCL(L)"            "SMP(L)"           
#>  [82] "CRE(L)"            "bL(L)"             "EBr3pw"           
#>  [85] "LX(L)"             "PRW"               "EPA(R)"           
#>  [88] "SNP(L)"            "CA(L)"             "LX(R)"            
#>  [91] "FBl4"              "MB(+ACA)(R)"       "FBl2"             
#>  [94] "PB(R2)"            "bL(R)"             "NO"               
#>  [97] "AL(R)"             "a'3(R)"            "SAD"              
#> [100] "b2(R)"             "EPA(L)"            "aL(R)"            
#> [103] "EBr5"              "IPS(R)"            "RUB(R)"           
#> [106] "a2(R)"             "PB(L8)"            "EBr2r4"           
#> [109] "PB(R9)"            "IB"                "g2(R)"            
#> [112] "FBl5"              "GOR(R)"            "PB(L5)"           
#> [115] "LAL(L)"            "GA(R)"             "FB"               
#> [118] "SPS(L)"            "VMNP"              "FBl6"             
#> [121] "AB(R)"             "PED(R)"            "PB(L7)"           
#> [124] "POC"               "CRE(-ROB,-RUB)(R)" "NO(R)"            
#> [127] "VES(R)"            "FBl9"              "GNG"              
#> [130] "GC"                "NO1(R)"            "ME(R)"            
#> [133] "AME(R)"            "VLNP(R)"           "vACA(R)"          
#> [136] "EBr6"              "SNP(R)"            "SMP(R)"           
#> [139] "LOP(R)"            "NO3(L)"            "FBl3"             
#> [142] "FBl7"              "PB(L4)"            "PB(R5)"           
#> [145] "AMMC"              "CRE(R)"            "SCL(R)"           
#> [148] "ROB(R)"            "LO(R)"             "EBr3am"           
#> [151] "NO2(L)"           
#> 
#> $`hemibrain:v1.0`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                      
#> [1,] "GET"  "/api/v:ver/npexplorer/autapses"          
#> [2,] "GET"  "/api/raw/keyvalue/key/:instance/:key"    
#> [3,] "POST" "/api/npexplorer/rankedtable"             
#> [4,] "POST" "/api/v:ver/npexplorer/commonconnectivity"
#> [5,] "GET"  "/api/v:ver/roimeshes/mesh/:dataset/:roi" 
#> [6,] "POST" "/api/v:ver/roimeshes/mesh/:dataset/:roi" 
#>      name                                                                                       
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/raw/keyvalue.masterAPI.getKV-fm"          
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getRankedTable-fm"
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                       

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }