Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn

optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...

methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()
#> $`fib19:v1.0`
#> $`fib19:v1.0`$`last-mod`
#> [1] "2018-08-04 00:48:21"
#> 
#> $`fib19:v1.0`$uuid
#> [1] "93f08942ac85495ab140d2808d893187"
#> 
#> $`fib19:v1.0`$ROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$superLevelROIs
#> [1] "ME"  "LO"  "LOP"
#> 
#> $`fib19:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`fib19:v1.0`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "AL-VP1l(R)"        "AL-VM2(R)"         "AL-DM5(R)"        
#>   [4] "NO"                "PB(R9)"            "PB(L7)"           
#>   [7] "NO1(L)"            "LO(R)"             "EPA(L)"           
#>  [10] "INP"               "bL(R)"             "AL-VA7m(R)"       
#>  [13] "PLP(R)"            "AL-VA1d(R)"        "PB(R1)"           
#>  [16] "GOR(R)"            "g1(R)"             "MB(+ACA)(R)"      
#>  [19] "AL-DL1(R)"         "AL-VL1(R)"         "b'L(R)"           
#>  [22] "OL(R)"             "AL-DA2(R)"         "AL-V(R)"          
#>  [25] "ME(R)"             "AL-DC3(R)"         "AL-DA1(R)"        
#>  [28] "AB(R)"             "AL-VM7v(R)"        "FBl2"             
#>  [31] "AL-DC4(L)"         "AL-VA7l(R)"        "ATL(L)"           
#>  [34] "GF(R)"             "AL-DM3(L)"         "g2(R)"            
#>  [37] "CAN(R)"            "PB(R6)"            "GOR(L)"           
#>  [40] "LAL(-GA)(R)"       "SCL(R)"            "IPS(R)"           
#>  [43] "AL-DL5(R)"         "VLNP(R)"           "CA(L)"            
#>  [46] "SAD"               "SPS(L)"            "vACA(R)"          
#>  [49] "PB(L5)"            "a2(R)"             "ROB(R)"           
#>  [52] "AL-DL2v(R)"        "EBr3d"             "gL(L)"            
#>  [55] "FBl9"              "AME(R)"            "dACA(R)"          
#>  [58] "a'3(R)"            "AL-DC1(L)"         "AL-DA2(L)"        
#>  [61] "b2(R)"             "BU(R)"             "AL-DC2(R)"        
#>  [64] "AL-DM1(R)"         "AL-DM2(R)"         "AL-VL2a(R)"       
#>  [67] "AL-VP1m(R)"        "PRW"               "NO2(R)"           
#>  [70] "MB(R)"             "LAL(L)"            "RUB(L)"           
#>  [73] "a3(R)"             "AL-DL3(R)"         "AL-VA6(R)"        
#>  [76] "b'1(R)"            "EB"                "AL-DC2(L)"        
#>  [79] "SNP(L)"            "AL-VP2(R)"         "AB(L)"            
#>  [82] "AMMC"              "SMP(L)"            "PB"               
#>  [85] "CRE(L)"            "FBl5"              "AL-VC1(R)"        
#>  [88] "AL-VC5(R)"         "PB(L2)"            "LOP(R)"           
#>  [91] "PB(L1)"            "CRE(-ROB,-RUB)(R)" "AL-DC4(R)"        
#>  [94] "AL-VM1(R)"         "gL(R)"             "AL-VM7d(L)"       
#>  [97] "VES(L)"            "AL-DA4m(R)"        "AL-VL2p(R)"       
#> [100] "AL-VM5v(R)"        "PED(R)"            "AL(R)"            
#> [103] "AL-VM4(R)"         "RUB(R)"            "SLP(R)"           
#> [106] "NO(L)"             "AL-VA6(L)"         "aL(L)"            
#> [109] "AL-DM3(R)"         "AL-VA1v(R)"        "AL-VC2(R)"        
#> [112] "bL(L)"             "CA(R)"             "AVLP(R)"          
#> [115] "AL-DA4l(R)"        "AL-DC1(R)"         "AL-DP1l(R)"       
#> [118] "mALT(L)"           "a'2(R)"            "FBl1"             
#> [121] "AL-VC4(R)"         "AL-VP5(R)"         "POC"              
#> [124] "EBr5"              "BU(L)"             "g3(R)"            
#> [127] "AL-VC3l(R)"        "PB(R3)"            "EBr1"             
#> [130] "PB(L8)"            "LX(L)"             "SIP(L)"           
#> [133] "SMP(R)"            "FBl3"              "AL-VM3(R)"        
#> [136] "NO(R)"             "AL-VA4(R)"         "AL-VC3m(R)"       
#> [139] "IB"                "EBr3am"            "AL-D(L)"          
#> [142] "AL-DL2d(R)"        "AL-DM6(L)"         "LX(R)"            
#> [145] "VES(R)"            "AL-DP1m(R)"        "WED(R)"           
#> [148] "AL-VM7v(L)"        "SAD(-AMMC)"        "FBl4"             
#> [151] "a'L(R)"            "EPA(R)"            "AL-DM5(L)"        
#> [154] "CRE(-RUB)(L)"      "FBl7"              "AL-VA5(R)"        
#> [157] "PB(R8)"            "ICL(R)"            "AL-D(R)"          
#> [160] "AL-DL4(L)"         "AL(L)"             "AL-DM2(L)"        
#> [163] "b1(R)"             "PB(R4)"            "AL-VA2(R)"        
#> [166] "SCL(L)"            "b'2(R)"            "AL-DA3(R)"        
#> [169] "VMNP"              "AL-DP1m(L)"        "PB(L4)"           
#> [172] "NO1(R)"            "NO3(L)"            "CX"               
#> [175] "g4(R)"             "EBr6"              "PB(R2)"           
#> [178] "lACA(R)"           "PB(L3)"            "PB(L6)"           
#> [181] "PENP"              "AL-VP3(R)"         "FLA(R)"           
#> [184] "AOT(R)"            "NO2(L)"            "AOTU(R)"          
#> [187] "AL-DL5(L)"         "AL-DM4(L)"         "AL-VP4(R)"        
#> [190] "SNP(R)"            "PB(R7)"            "GC"               
#> [193] "a'1(R)"            "FB-column3"        "ATL(R)"           
#> [196] "LAL(R)"            "SPS(R)"            "AL-VM7d(R)"       
#> [199] "g5(R)"             "a1(R)"             "MB(L)"            
#> [202] "PB(R5)"            "AL-VM5d(R)"        "PVLP(R)"          
#> [205] "FBl8"              "AL-DM6(R)"         "AL-DM1(L)"        
#> [208] "NO3(R)"            "CRE(R)"            "b'L(L)"           
#> [211] "GA(R)"             "EBr2r4"            "AL-DA3(L)"        
#> [214] "PB(L9)"            "FBl6"              "LH(R)"            
#> [217] "a'L(L)"            "AL-DL4(R)"         "mALT(R)"          
#> [220] "AL-VA3(R)"         "ICL(L)"            "AL-VP1d(R)"       
#> [223] "SIP(R)"            "aL(R)"             "FB"               
#> [226] "AL-DA4m(L)"        "EBr3pw"            "IB112"            
#> [229] "GNG"               "AL-DM4(R)"         "hemibrain"        
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PED(R)"            "LX(R)"             "VES(R)"           
#>   [4] "RUB(R)"            "a3(R)"             "PLP(R)"           
#>   [7] "EB"                "PB(L6)"            "EPA(L)"           
#>  [10] "mALT(R)"           "PB(L9)"            "PB(L4)"           
#>  [13] "CA(L)"             "b'1(R)"            "NO2(L)"           
#>  [16] "LAL(L)"            "a'2(R)"            "NO2(R)"           
#>  [19] "RUB(L)"            "BU(L)"             "bL(L)"            
#>  [22] "EBr3am"            "SMP(L)"            "NO1(L)"           
#>  [25] "EBr1"              "SLP(R)"            "NO3(L)"           
#>  [28] "SAD(-AMMC)"        "VES(L)"            "IB"               
#>  [31] "LAL(-GA)(R)"       "PB(R8)"            "a'L(L)"           
#>  [34] "EBr3d"             "PB(L7)"            "b'L(L)"           
#>  [37] "AB(L)"             "aL(L)"             "PB(R5)"           
#>  [40] "SAD"               "CA(R)"             "AME(R)"           
#>  [43] "CAN(R)"            "PB(R2)"            "INP"              
#>  [46] "VLNP(R)"           "GOR(R)"            "CX"               
#>  [49] "PB(R7)"            "NO3(R)"            "EBr5"             
#>  [52] "FBl7"              "EPA(R)"            "ICL(R)"           
#>  [55] "NO(L)"             "AMMC"              "PB(L3)"           
#>  [58] "lACA(R)"           "SCL(R)"            "FBl1"             
#>  [61] "SCL(L)"            "a'3(R)"            "ATL(L)"           
#>  [64] "GC"                "BU(R)"             "LH(R)"            
#>  [67] "b1(R)"             "FBl5"              "GNG"              
#>  [70] "vACA(R)"           "GF(R)"             "g4(R)"            
#>  [73] "CRE(L)"            "FBl3"              "PVLP(R)"          
#>  [76] "FBl4"              "PB(R3)"            "PENP"             
#>  [79] "g3(R)"             "GA(R)"             "FB"               
#>  [82] "PB(L2)"            "AOTU(R)"           "GOR(L)"           
#>  [85] "MB(L)"             "PB(L8)"            "IPS(R)"           
#>  [88] "bL(R)"             "AL(R)"             "AL(L)"            
#>  [91] "SIP(R)"            "NO"                "AB(R)"            
#>  [94] "MB(+ACA)(R)"       "LX(L)"             "CRE(-ROB,-RUB)(R)"
#>  [97] "g1(R)"             "NO(R)"             "gL(R)"            
#> [100] "SNP(L)"            "ROB(R)"            "g5(R)"            
#> [103] "FBl6"              "PB(R4)"            "b'2(R)"           
#> [106] "a2(R)"             "SPS(R)"            "EBr6"             
#> [109] "SNP(R)"            "OL(R)"             "FBl8"             
#> [112] "FBl9"              "ATL(R)"            "PB(R1)"           
#> [115] "ICL(L)"            "LO(R)"             "FBl2"             
#> [118] "SIP(L)"            "aL(R)"             "LAL(R)"           
#> [121] "SMP(R)"            "g2(R)"             "FLA(R)"           
#> [124] "AVLP(R)"           "PB(L5)"            "hemibrain"        
#> [127] "PRW"               "CRE(R)"            "b'L(R)"           
#> [130] "a1(R)"             "PB(R6)"            "dACA(R)"          
#> [133] "a'1(R)"            "PB(L1)"            "ME(R)"            
#> [136] "a'L(R)"            "EBr2r4"            "POC"              
#> [139] "mALT(L)"           "b2(R)"             "NO1(R)"           
#> [142] "PB"                "LOP(R)"            "EBr3pw"           
#> [145] "MB(R)"             "WED(R)"            "SPS(L)"           
#> [148] "VMNP"              "gL(L)"             "PB(R9)"           
#> [151] "AOT(R)"           
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "ICL(L)"            "b'L(L)"            "AL-DA4m(R)"       
#>   [4] "AL-VP1d(R)"        "LOP(R)"            "CRE(L)"           
#>   [7] "AVLP(R)"           "AL-VP2(R)"         "AL-DM5(R)"        
#>  [10] "FBl4"              "FBl5"              "LH(R)"            
#>  [13] "AL-DM2(R)"         "CRE(-ROB,-RUB)(R)" "FLA(R)"           
#>  [16] "AL-D(L)"           "AL-DM6(L)"         "EBr3am"           
#>  [19] "CX"                "hemibrain"         "OL(R)"            
#>  [22] "AL-VA7m(R)"        "a'2(R)"            "a'L(R)"           
#>  [25] "PB(R2)"            "AL-DL4(R)"         "AL-VL2p(R)"       
#>  [28] "b'L(R)"            "LAL(R)"            "BU(R)"            
#>  [31] "AL(L)"             "GOR(L)"            "AL-DL5(L)"        
#>  [34] "ROB(R)"            "a'1(R)"            "RUB(L)"           
#>  [37] "g4(R)"             "MB(R)"             "CA(L)"            
#>  [40] "AL-DP1m(L)"        "CRE(-RUB)(L)"      "AL-VA5(R)"        
#>  [43] "LAL(L)"            "FBl2"              "AL-VA4(R)"        
#>  [46] "PENP"              "PB(L6)"            "INP"              
#>  [49] "PB(R6)"            "EB"                "GNG"              
#>  [52] "GOR(R)"            "AL-VL1(R)"         "GC"               
#>  [55] "g3(R)"             "g2(R)"             "AL-DM1(R)"        
#>  [58] "AL-D(R)"           "AL-VA1d(R)"        "GF(R)"            
#>  [61] "FBl8"              "AME(R)"            "vACA(R)"          
#>  [64] "SAD"               "AL-DC3(R)"         "PB(L1)"           
#>  [67] "CAN(R)"            "EBr3d"             "NO"               
#>  [70] "PB(R4)"            "AL-VP1l(R)"        "AB(L)"            
#>  [73] "b'1(R)"            "PB(L2)"            "AL-DM1(L)"        
#>  [76] "a'L(L)"            "AL-VC2(R)"         "AL-VC3m(R)"       
#>  [79] "PB"                "EBr3pw"            "FBl9"             
#>  [82] "AL-VA6(R)"         "RUB(R)"            "CRE(R)"           
#>  [85] "WED(R)"            "AL-DM3(L)"         "PB(L3)"           
#>  [88] "MB(+ACA)(R)"       "dACA(R)"           "EBr2r4"           
#>  [91] "NO3(R)"            "LO(R)"             "AL-VA3(R)"        
#>  [94] "AL-VM7v(R)"        "ATL(L)"            "AMMC"             
#>  [97] "gL(R)"             "FB"                "ME(R)"            
#> [100] "AL-VM3(R)"         "mALT(R)"           "PB(L4)"           
#> [103] "AL-DA4m(L)"        "SCL(R)"            "AL-DA4l(R)"       
#> [106] "b2(R)"             "AL-DM2(L)"         "NO3(L)"           
#> [109] "a'3(R)"            "MB(L)"             "PVLP(R)"          
#> [112] "PB(R3)"            "AL-DL5(R)"         "AL-VM1(R)"        
#> [115] "VLNP(R)"           "SIP(R)"            "EBr1"             
#> [118] "SPS(L)"            "bL(L)"             "NO(L)"            
#> [121] "AL-VA1v(R)"        "PED(R)"            "EPA(R)"           
#> [124] "PB(R8)"            "AL-VA7l(R)"        "CA(R)"            
#> [127] "AL-DL3(R)"         "aL(L)"             "PB(L7)"           
#> [130] "PB(R1)"            "PB(L9)"            "IPS(R)"           
#> [133] "EPA(L)"            "AL-DC1(R)"         "AL-DP1m(R)"       
#> [136] "AL-DM4(L)"         "BU(L)"             "AL-DA2(L)"        
#> [139] "EBr5"              "NO(R)"             "bL(R)"            
#> [142] "AL-DA1(R)"         "AOT(R)"            "PB(L8)"           
#> [145] "b1(R)"             "FBl7"              "VES(L)"           
#> [148] "gL(L)"             "AL-DC1(L)"         "PB(R9)"           
#> [151] "AL-DA3(L)"         "ICL(R)"            "AL-VM2(R)"        
#> [154] "PB(L5)"            "a1(R)"             "SAD(-AMMC)"       
#> [157] "AL-VM4(R)"         "mALT(L)"           "ATL(R)"           
#> [160] "g1(R)"             "VES(R)"            "FBl6"             
#> [163] "AL-DC4(R)"         "AL-VC5(R)"         "PRW"              
#> [166] "LAL(-GA)(R)"       "NO1(L)"            "AL-DL4(L)"        
#> [169] "AL-VC3l(R)"        "AL-VM7d(R)"        "IB"               
#> [172] "b'2(R)"            "GA(R)"             "FBl3"             
#> [175] "AL-VM5v(R)"        "AOTU(R)"           "AL-DP1l(R)"       
#> [178] "AL-VP4(R)"         "PB(R7)"            "AL-DC2(L)"        
#> [181] "AL-VA6(L)"         "AL-DL2d(R)"        "AL-VL2a(R)"       
#> [184] "PLP(R)"            "aL(R)"             "NO1(R)"           
#> [187] "LX(L)"             "AL(R)"             "AL-DM3(R)"        
#> [190] "AL-DM4(R)"         "SMP(R)"            "SIP(L)"           
#> [193] "AL-VP5(R)"         "AL-VC4(R)"         "SPS(R)"           
#> [196] "SNP(R)"            "SLP(R)"            "AL-DA3(R)"        
#> [199] "a2(R)"             "a3(R)"             "AL-DA2(R)"        
#> [202] "SMP(L)"            "AB(R)"             "AL-VP1m(R)"       
#> [205] "AL-VM7d(L)"        "AL-VM7v(L)"        "AL-DM6(R)"        
#> [208] "AL-VM5d(R)"        "POC"               "SCL(L)"           
#> [211] "AL-DC4(L)"         "AL-VA2(R)"         "AL-DC2(R)"        
#> [214] "FBl1"              "AL-V(R)"           "AL-VP3(R)"        
#> [217] "PB(R5)"            "LX(R)"             "g5(R)"            
#> [220] "NO2(L)"            "AL-DM5(L)"         "FB-column3"       
#> [223] "lACA(R)"           "EBr6"              "AL-DL2v(R)"       
#> [226] "VMNP"              "NO2(R)"            "AL-DL1(R)"        
#> [229] "AL-VC1(R)"         "SNP(L)"           
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "aL(R)"             "AL-VM1(R)"         "IB"               
#>   [4] "AL-VM7v(L)"        "INP"               "AL-DC4(R)"        
#>   [7] "AL(L)"             "AL-D(L)"           "AL-DA1(R)"        
#>  [10] "a'3(R)"            "vACA(R)"           "PRW"              
#>  [13] "AL-VL2p(R)"        "AL-VM5v(R)"        "AL-DM6(R)"        
#>  [16] "AL-VM3(R)"         "AL-V(R)"           "b'2(R)"           
#>  [19] "AOTU(R)"           "AL-DC2(R)"         "gL(R)"            
#>  [22] "AL-DC1(R)"         "g4(R)"             "AL-DP1m(L)"       
#>  [25] "AL-VM7d(L)"        "AL-VP1d(R)"        "FLA(R)"           
#>  [28] "CRE(-RUB)(L)"      "PB(R4)"            "SCL(L)"           
#>  [31] "NO3(L)"            "EBr3pw"            "AB(R)"            
#>  [34] "FBl2"              "MB(R)"             "AL-VP2(R)"        
#>  [37] "AL-DM3(L)"         "SAD(-AMMC)"        "FBl7"             
#>  [40] "VES(L)"            "AL-DA2(R)"         "GC"               
#>  [43] "CRE(R)"            "SPS(L)"            "AL-VA6(R)"        
#>  [46] "FBl6"              "LX(L)"             "b'L(L)"           
#>  [49] "SAD"               "CAN(R)"            "AL-VP5(R)"        
#>  [52] "AL-DM2(L)"         "OL(R)"             "AL-DM2(R)"        
#>  [55] "AL-DC2(L)"         "bL(L)"             "AL-VC4(R)"        
#>  [58] "AL-DA3(L)"         "FBl3"              "FBl8"             
#>  [61] "AL-VC2(R)"         "RUB(R)"            "AL-VC3l(R)"       
#>  [64] "a2(R)"             "MB(L)"             "EPA(L)"           
#>  [67] "GOR(L)"            "CA(R)"             "ATL(L)"           
#>  [70] "FBl5"              "PB(R7)"            "SCL(R)"           
#>  [73] "EBr3d"             "SNP(R)"            "AL-VM2(R)"        
#>  [76] "mALT(R)"           "g2(R)"             "PB(R9)"           
#>  [79] "b2(R)"             "VMNP"              "AL-VP1l(R)"       
#>  [82] "LH(R)"             "a'L(L)"            "g3(R)"            
#>  [85] "AL-DM5(L)"         "SNP(L)"            "aL(L)"            
#>  [88] "AL-VC5(R)"         "NO3(R)"            "g5(R)"            
#>  [91] "FBl1"              "AL-VM4(R)"         "BU(L)"            
#>  [94] "FBl4"              "NO2(R)"            "a3(R)"            
#>  [97] "AL-DM3(R)"         "g1(R)"             "AL-VA6(L)"        
#> [100] "b1(R)"             "EPA(R)"            "GNG"              
#> [103] "PENP"              "WED(R)"            "a'L(R)"           
#> [106] "ATL(R)"            "AL-DM1(R)"         "AL-VA7l(R)"       
#> [109] "AL-VM7v(R)"        "VLNP(R)"           "PB(L8)"           
#> [112] "PB(R8)"            "MB(+ACA)(R)"       "POC"              
#> [115] "PB(L2)"            "FB"                "ICL(R)"           
#> [118] "AL-VP4(R)"         "PB(R3)"            "AVLP(R)"          
#> [121] "IPS(R)"            "AL-VL2a(R)"        "mALT(L)"          
#> [124] "GF(R)"             "NO1(L)"            "a'2(R)"           
#> [127] "AL-D(R)"           "AL-VA5(R)"         "AL-DL4(R)"        
#> [130] "PED(R)"            "NO"                "GOR(R)"           
#> [133] "AOT(R)"            "EBr5"              "PB(L1)"           
#> [136] "AL-DM5(R)"         "PB(L3)"            "CRE(L)"           
#> [139] "AL-VC3m(R)"        "AL-VM7d(R)"        "PB(L5)"           
#> [142] "GA(R)"             "NO1(R)"            "PVLP(R)"          
#> [145] "LAL(-GA)(R)"       "AL-DL1(R)"         "PB(R2)"           
#> [148] "b'1(R)"            "AL-DA4m(R)"        "AL-VA1v(R)"       
#> [151] "AL-VA3(R)"         "dACA(R)"           "NO(L)"            
#> [154] "LAL(R)"            "AL-DA3(R)"         "PB(R5)"           
#> [157] "EBr3am"            "SIP(L)"            "AL-DC1(L)"        
#> [160] "FBl9"              "ICL(L)"            "AL(R)"            
#> [163] "AL-DA4l(R)"        "AL-VM5d(R)"        "LX(R)"            
#> [166] "EB"                "NO(R)"             "hemibrain"        
#> [169] "EBr2r4"            "AL-VA4(R)"         "EBr1"             
#> [172] "PLP(R)"            "PB(L6)"            "SMP(L)"           
#> [175] "ME(R)"             "AL-DA4m(L)"        "AL-DM4(L)"        
#> [178] "AL-VP3(R)"         "AL-VP1m(R)"        "AL-DL4(L)"        
#> [181] "AL-DL5(R)"         "AL-DP1l(R)"        "AL-DM1(L)"        
#> [184] "PB(R1)"            "AL-DL3(R)"         "bL(R)"            
#> [187] "AL-VA2(R)"         "AL-DA2(L)"         "SLP(R)"           
#> [190] "VES(R)"            "AL-DM4(R)"         "lACA(R)"          
#> [193] "b'L(R)"            "AL-DM6(L)"         "AL-DC4(L)"        
#> [196] "CRE(-ROB,-RUB)(R)" "AL-DL2v(R)"        "AL-DP1m(R)"       
#> [199] "AL-VA1d(R)"        "LOP(R)"            "LAL(L)"           
#> [202] "PB(L9)"            "AME(R)"            "SPS(R)"           
#> [205] "AL-VA7m(R)"        "NO2(L)"            "LO(R)"            
#> [208] "SMP(R)"            "a1(R)"             "FB-column3"       
#> [211] "AL-DL5(L)"         "AL-DL2d(R)"        "AMMC"             
#> [214] "PB(R6)"            "EBr6"              "gL(L)"            
#> [217] "AL-VL1(R)"         "CA(L)"             "SIP(R)"           
#> [220] "a'1(R)"            "AL-VC1(R)"         "BU(R)"            
#> [223] "RUB(L)"            "PB"                "ROB(R)"           
#> [226] "AB(L)"             "CX"                "PB(L7)"           
#> [229] "PB(L4)"            "AL-DC3(R)"        
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 
# }
# \donttest{
# list API endpoints
head(neuprint_available())
#>      method path                                          
#> [1,] "POST" "/api/npexplorer/celltype/:dataset/:type"     
#> [2,] "GET"  "/api/v:ver/dbmeta/version"                   
#> [3,] "POST" "/logout"                                     
#> [4,] "GET"  "/token"                                      
#> [5,] "POST" "/api/v:ver/raw/cypher/transaction/:id/commit"
#> [6,] "GET"  "/api/dbmeta/version"                         
#>      name                                                                                    
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getCellType-fm"
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [3,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                
#> [4,] "github.com/janelia-flyem/echo-secure.EchoSecure.AuthMiddleware.func1.1"                
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                    
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta.storeAPI.getVersion-fm"         
# }
# \donttest{
neuprint_version()
#> $Version
#> [1] "0.1.0"
#> 
# }