Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server see ?neuprint_login. If NULL, your defaults set in your R.profile or R.environ are used.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v1.0`
#> $`hemibrain:v1.0`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0`$ROIs
#>   [1] "hemibrain"         "LAL(-GA)(R)"       "bL(L)"            
#>   [4] "LH(R)"             "GNG"               "SCL(R)"           
#>   [7] "EPA(R)"            "FBl2"              "g1(R)"            
#>  [10] "BU(L)"             "PB(L3)"            "PB(L1)"           
#>  [13] "SIP(R)"            "PVLP(R)"           "LX(L)"            
#>  [16] "g2(R)"             "PB(R8)"            "VMNP"             
#>  [19] "g4(R)"             "EBr3pw"            "PB(L7)"           
#>  [22] "POC"               "AOT(R)"            "LOP(R)"           
#>  [25] "SIP(L)"            "SPS(L)"            "CRE(L)"           
#>  [28] "LAL(R)"            "CRE(-ROB,-RUB)(R)" "NO2(L)"           
#>  [31] "gL(R)"             "CA(R)"             "VLNP(R)"          
#>  [34] "WED(R)"            "PB(L8)"            "NO"               
#>  [37] "FBl9"              "PB(R3)"            "b'2(R)"           
#>  [40] "EBr1"              "NO2(R)"            "CRE(R)"           
#>  [43] "AL(L)"             "SNP(L)"            "FBl1"             
#>  [46] "BU(R)"             "SCL(L)"            "SAD"              
#>  [49] "b'L(R)"            "PB(R9)"            "b1(R)"            
#>  [52] "AB(L)"             "SMP(L)"            "PLP(R)"           
#>  [55] "CA(L)"             "EBr6"              "FBl7"             
#>  [58] "CAN(R)"            "GOR(R)"            "SLP(R)"           
#>  [61] "SMP(R)"            "PED(R)"            "EBr2r4"           
#>  [64] "gL(L)"             "PB(R2)"            "NO(L)"            
#>  [67] "b'1(R)"            "RUB(L)"            "a'L(L)"           
#>  [70] "PB(R7)"            "EBr5"              "ME(R)"            
#>  [73] "SNP(R)"            "ROB(R)"            "LO(R)"            
#>  [76] "AL(R)"             "EBr3am"            "vACA(R)"          
#>  [79] "CX"                "FBl3"              "PB(R4)"           
#>  [82] "PENP"              "a2(R)"             "NO1(L)"           
#>  [85] "dACA(R)"           "PB(R1)"            "SAD(-AMMC)"       
#>  [88] "PB(L4)"            "RUB(R)"            "a'3(R)"           
#>  [91] "GA(R)"             "IB"                "ATL(L)"           
#>  [94] "FB"                "NO(R)"             "VES(R)"           
#>  [97] "b'L(L)"            "OL(R)"             "g3(R)"            
#> [100] "a'2(R)"            "PB(L6)"            "MB(+ACA)(R)"      
#> [103] "FBl5"              "bL(R)"             "PB(R6)"           
#> [106] "a'1(R)"            "NO3(L)"            "FBl4"             
#> [109] "INP"               "AOTU(R)"           "PB"               
#> [112] "lACA(R)"           "NO3(R)"            "EB"               
#> [115] "ICL(L)"            "LAL(L)"            "GF(R)"            
#> [118] "aL(R)"             "MB(L)"             "NO1(R)"           
#> [121] "AMMC"              "AVLP(R)"           "PB(L2)"           
#> [124] "SPS(R)"            "GOR(L)"            "AME(R)"           
#> [127] "PB(R5)"            "FLA(R)"            "a'L(R)"           
#> [130] "IPS(R)"            "aL(L)"             "VES(L)"           
#> [133] "MB(R)"             "g5(R)"             "ICL(R)"           
#> [136] "a3(R)"             "mALT(R)"           "GC"               
#> [139] "FBl6"              "mALT(L)"           "LX(R)"            
#> [142] "PB(L9)"            "FBl8"              "b2(R)"            
#> [145] "a1(R)"             "PRW"               "EBr3d"            
#> [148] "ATL(R)"            "AB(R)"             "EPA(L)"           
#> [151] "PB(L5)"           
#> 
#> $`hemibrain:v1.0`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method path                                        
#> [1,] "POST" "/api/v:ver/raw/cypher/transaction/:id/kill"
#> [2,] "GET"  "/api/v:ver/cached/dailytype"               
#> [3,] "GET"  "/api/v:ver/dbmeta/datasets"                
#> [4,] "GET"  "/api/npexplorer/roiconnectivity"           
#> [5,] "GET"  "/api/npexplorer/rankedtable"               
#> [6,] "GET"  "/oauth2callback"                           
#>      name                                                                                           
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [3,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                           
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getROIConnectivity-fm"
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getRankedTable-fm"    
#> [6,] "github.com/janelia-flyem/echo-secure.oauthCallbackHandler"                                    

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }