Find out what some information about your neuPrint server

Get summary information about the datasets hosted by the neuPrint server in which you are interested

neuprint_datasets(conn = NULL, ...)

neuprint_database(conn = NULL, ...)

neuprint_available(conn = NULL, ...)

neuprint_version(conn = NULL, ...)

Arguments

conn
optional, a neuprintr connection object, which also specifies the neuPrint server. If NULL, the defaults set in your .Rprofile or .Renviron are used. See neuprint_login for details.

...
methods passed to neuprint_login

See also

Examples

# \donttest{
# list details for all the available datasets
neuprint_datasets()

#> $`hemibrain:v0.9`
#> $`hemibrain:v0.9`$`last-mod`
#> [1] "2019-06-27 18:08:56"
#> 
#> $`hemibrain:v0.9`$uuid
#> [1] "569625b31af845479fe4f13ff53f868e"
#> 
#> $`hemibrain:v0.9`$ROIs
#>   [1] "PB"                "FB-column3"        "FBl8"             
#>   [4] "AL-VM1(R)"         "MB(L)"             "EB"               
#>   [7] "PB(R1)"            "FB"                "PB(R3)"           
#>  [10] "INP"               "AL-VM3(R)"         "PVLP(R)"          
#>  [13] "AL-VL1(R)"         "AL-VL2p(R)"        "SIP(L)"           
#>  [16] "AL-DC1(L)"         "AL-DM2(R)"         "AL-VL2a(R)"       
#>  [19] "AL-VM5v(R)"        "AL-VP5(R)"         "SMP(L)"           
#>  [22] "a'L(R)"            "PB(R9)"            "gL(R)"            
#>  [25] "AL-DC1(R)"         "AL-VA3(R)"         "AOT(R)"           
#>  [28] "PB(R5)"            "PB(L3)"            "SIP(R)"           
#>  [31] "AL-DM6(R)"         "AL-VP3(R)"         "SPS(R)"           
#>  [34] "PB(R8)"            "AL-DL2d(R)"        "AL-DM1(R)"        
#>  [37] "AL-DM6(L)"         "PB(L9)"            "a'L(L)"           
#>  [40] "mALT(R)"           "VMNP"              "SNP(L)"           
#>  [43] "g2(R)"             "dACA(R)"           "AL-DL5(L)"        
#>  [46] "AL-D(L)"           "VES(R)"            "AL-DA3(R)"        
#>  [49] "AL-DM3(R)"         "AL-VC4(R)"         "SMP(R)"           
#>  [52] "EBr3am"            "AL-VA7m(R)"        "POC"              
#>  [55] "NO2(L)"            "a'2(R)"            "GOR(L)"           
#>  [58] "RUB(L)"            "SLP(R)"            "AL-DM5(R)"        
#>  [61] "AL-VA7l(R)"        "LAL(L)"            "a2(R)"            
#>  [64] "GA(R)"             "b2(R)"             "AL-DC4(L)"        
#>  [67] "AVLP(R)"           "PB(L5)"            "ATL(L)"           
#>  [70] "BU(R)"             "PB(L1)"            "AL-DL2v(R)"       
#>  [73] "g4(R)"             "FBl4"              "CRE(-ROB,-RUB)(R)"
#>  [76] "AL-DM4(L)"         "b'L(L)"            "EPA(R)"           
#>  [79] "AL-DL1(R)"         "AL-VP2(R)"         "VLNP(R)"          
#>  [82] "SNP(R)"            "aL(L)"             "AL-DL3(R)"        
#>  [85] "RUB(R)"            "PB(L6)"            "a'1(R)"           
#>  [88] "LH(R)"             "AL-VC3m(R)"        "AL-VM5d(R)"       
#>  [91] "PB(L8)"            "gL(L)"             "AL-DA4m(L)"       
#>  [94] "bL(L)"             "LX(L)"             "AOTU(R)"          
#>  [97] "FLA(R)"            "ICL(R)"            "AL-VP4(R)"        
#> [100] "EBr2r4"            "GNG"               "VES(L)"           
#> [103] "PB(R6)"            "aL(R)"             "AL-DC2(L)"        
#> [106] "AMMC"              "EBr1"              "AL-DA3(L)"        
#> [109] "AL-DM3(L)"         "NO(R)"             "FBl7"             
#> [112] "AL-VA1d(R)"        "g1(R)"             "IB"               
#> [115] "SCL(L)"            "PB(L2)"            "NO1(L)"           
#> [118] "NO(L)"             "AL-DP1m(R)"        "ATL(R)"           
#> [121] "GF(R)"             "WED(R)"            "SCL(R)"           
#> [124] "AL-VA6(L)"         "NO"                "MB(+ACA)(R)"      
#> [127] "OL(R)"             "LAL(-GA)(R)"       "LX(R)"            
#> [130] "AL-VM7d(L)"        "EBr3pw"            "LO(R)"            
#> [133] "EPA(L)"            "PB(L4)"            "b'2(R)"           
#> [136] "LOP(R)"            "hemibrain"         "EBr6"             
#> [139] "PB(L7)"            "AL-DL5(R)"         "PENP"             
#> [142] "g3(R)"             "AL(L)"             "g5(R)"            
#> [145] "EBr5"              "AL-DA2(L)"         "FBl1"             
#> [148] "AL-DM4(R)"         "NO3(L)"            "SAD(-AMMC)"       
#> [151] "AL-VC1(R)"         "AL-DM1(L)"         "PB(R4)"           
#> [154] "MB(R)"             "AL-DC4(R)"         "AL-D(R)"          
#> [157] "AL-VM4(R)"         "lACA(R)"           "a3(R)"            
#> [160] "a'3(R)"            "AL-VM2(R)"         "SPS(L)"           
#> [163] "ROB(R)"            "FBl2"              "AL-DM2(L)"        
#> [166] "CRE(-RUB)(L)"      "AB(R)"             "AL-VA2(R)"        
#> [169] "CX"                "CA(L)"             "LAL(R)"           
#> [172] "b'1(R)"            "PED(R)"            "ME(R)"            
#> [175] "AL-DL4(L)"         "FBl6"              "CAN(R)"           
#> [178] "CA(R)"             "NO2(R)"            "GOR(R)"           
#> [181] "AL-DA2(R)"         "AL-VA1v(R)"        "PB(R7)"           
#> [184] "AL-VC3l(R)"        "ICL(L)"            "FBl9"             
#> [187] "AL-V(R)"           "AL-VM7d(R)"        "CRE(R)"           
#> [190] "BU(L)"             "PLP(R)"            "AL-VM7v(L)"       
#> [193] "AME(R)"            "AL-DP1l(R)"        "AL-DP1m(L)"       
#> [196] "AL-VC2(R)"         "SAD"               "AL(R)"            
#> [199] "AL-VP1l(R)"        "bL(R)"             "AL-DC3(R)"        
#> [202] "AL-VP1m(R)"        "b'L(R)"            "CRE(L)"           
#> [205] "FBl5"              "PB(R2)"            "b1(R)"            
#> [208] "AL-DC2(R)"         "AL-VA6(R)"         "AB(L)"            
#> [211] "vACA(R)"           "FBl3"              "IPS(R)"           
#> [214] "AL-DA4l(R)"        "AL-VA4(R)"         "PRW"              
#> [217] "NO1(R)"            "AL-DA1(R)"         "AL-DL4(R)"        
#> [220] "AL-VA5(R)"         "AL-VC5(R)"         "AL-VM7v(R)"       
#> [223] "a1(R)"             "EBr3d"             "NO3(R)"           
#> [226] "AL-DM5(L)"         "mALT(L)"           "GC"               
#> [229] "IB112"             "AL-DA4m(R)"        "AL-VP1d(R)"       
#> 
#> $`hemibrain:v0.9`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v0.9`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v0.9`$hidden
#> [1] TRUE
#> 
#> 
#> $`hemibrain:v1.0.1`
#> $`hemibrain:v1.0.1`$`last-mod`
#> [1] "2019-12-20T23:11:02"
#> 
#> $`hemibrain:v1.0.1`$uuid
#> [1] "52a13328874c4bb7b15dc4280da26576"
#> 
#> $`hemibrain:v1.0.1`$ROIs
#>   [1] "PENP"              "NO1(L)"            "PB(R9)"           
#>   [4] "FBl9"              "b'L(L)"            "POC"              
#>   [7] "PB(R6)"            "lACA(R)"           "EBr1"             
#>  [10] "BU(L)"             "PB(L1)"            "CRE(L)"           
#>  [13] "NO2(R)"            "SMP(L)"            "a2(R)"            
#>  [16] "GA(R)"             "aL(R)"             "PB(R8)"           
#>  [19] "g4(R)"             "SMP(R)"            "PB(L2)"           
#>  [22] "EBr3am"            "FLA(R)"            "FBl8"             
#>  [25] "AL(R)"             "vACA(R)"           "RUB(L)"           
#>  [28] "RUB(R)"            "EBr2r4"            "PB(R1)"           
#>  [31] "FBl3"              "a3(R)"             "EBr6"             
#>  [34] "PB(L6)"            "PB(R5)"            "b'1(R)"           
#>  [37] "PB(R7)"            "EB"                "hemibrain"        
#>  [40] "PRW"               "bL(L)"             "MB(+ACA)(R)"      
#>  [43] "b1(R)"             "a'L(L)"            "GOR(L)"           
#>  [46] "AVLP(R)"           "LX(L)"             "LOP(R)"           
#>  [49] "g3(R)"             "ROB(R)"            "LO(R)"            
#>  [52] "EPA(L)"            "LAL(R)"            "OL(R)"            
#>  [55] "GC"                "PVLP(R)"           "PB(L4)"           
#>  [58] "PB(L5)"            "SCL(R)"            "dACA(R)"          
#>  [61] "a'2(R)"            "FBl6"              "AOT(R)"           
#>  [64] "BU(R)"             "CA(R)"             "PB(R4)"           
#>  [67] "CRE(-ROB,-RUB)(R)" "AOTU(R)"           "FB"               
#>  [70] "PB"                "EBr3d"             "bL(R)"            
#>  [73] "ATL(L)"            "WED(R)"            "a'1(R)"           
#>  [76] "FBl7"              "PB(R2)"            "AB(R)"            
#>  [79] "SIP(L)"            "SPS(L)"            "gL(L)"            
#>  [82] "SIP(R)"            "PB(L9)"            "SNP(L)"           
#>  [85] "ICL(R)"            "SCL(L)"            "LAL(L)"           
#>  [88] "LAL(-GA)(R)"       "NO1(R)"            "SAD(-AMMC)"       
#>  [91] "FBl2"              "EBr5"              "SNP(R)"           
#>  [94] "EBr3pw"            "ICL(L)"            "a1(R)"            
#>  [97] "SAD"               "VMNP"              "AL(L)"            
#> [100] "FBl1"              "FBl4"              "VES(L)"           
#> [103] "SPS(R)"            "g5(R)"             "VLNP(R)"          
#> [106] "b'2(R)"            "b2(R)"             "EPA(R)"           
#> [109] "NO(L)"             "GF(R)"             "mALT(R)"          
#> [112] "g1(R)"             "NO3(R)"            "PB(L7)"           
#> [115] "AME(R)"            "PB(L3)"            "PED(R)"           
#> [118] "b'L(R)"            "NO(R)"             "LH(R)"            
#> [121] "CAN(R)"            "aL(L)"             "MB(L)"            
#> [124] "a'L(R)"            "FBl5"              "LX(R)"            
#> [127] "ME(R)"             "gL(R)"             "PB(R3)"           
#> [130] "SLP(R)"            "a'3(R)"            "AMMC"             
#> [133] "NO2(L)"            "NO"                "GNG"              
#> [136] "MB(R)"             "mALT(L)"           "CRE(R)"           
#> [139] "g2(R)"             "PB(L8)"            "PLP(R)"           
#> [142] "VES(R)"            "IPS(R)"            "GOR(R)"           
#> [145] "CA(L)"             "NO3(L)"            "ATL(R)"           
#> [148] "AB(L)"             "CX"                "INP"              
#> [151] "IB"               
#> 
#> $`hemibrain:v1.0.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.0.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.0.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.1`
#> $`hemibrain:v1.1`$`last-mod`
#> [1] "2020-05-30 10:36:36"
#> 
#> $`hemibrain:v1.1`$uuid
#> [1] "20631f94c3f446d7864bc55bf515706e"
#> 
#> $`hemibrain:v1.1`$ROIs
#>   [1] "AL-DL1(R)"         "AL-DM5(R)"         "AOT(R)"           
#>   [4] "mALT(R)"           "ME(R)"             "gL(L)"            
#>   [7] "g4(R)"             "FBl5"              "SPS(L)"           
#>  [10] "PB"                "AL-VA3(R)"         "ICL(R)"           
#>  [13] "AL-VC1(R)"         "AL-VC3l(R)"        "g3(R)"            
#>  [16] "AL-DL4(L)"         "a'L(L)"            "EBr3d"            
#>  [19] "ICL(L)"            "AL-VA7l(R)"        "AL-VC3m(R)"       
#>  [22] "b'2(R)"            "g5(R)"             "AL-VM7d(L)"       
#>  [25] "AB(L)"             "GF(R)"             "CAN(R)"           
#>  [28] "PB(L9)"            "g2(R)"             "AL-VM2(R)"        
#>  [31] "AL-DA3(L)"         "AL-DC3(R)"         "AL-DL2v(R)"       
#>  [34] "a2(R)"             "EBr2r4"            "EBr3pw"           
#>  [37] "AL-VA1v(R)"        "LAL(-GA)(R)"       "EBr6"             
#>  [40] "AL-D(L)"           "SIP(R)"            "AL-DM5(L)"        
#>  [43] "SIP(L)"            "AL-DA2(L)"         "AL-VP2(R)"        
#>  [46] "EPA(R)"            "NO(R)"             "RUB(L)"           
#>  [49] "a1(R)"             "SAD"               "GOR(L)"           
#>  [52] "PB(R3)"            "AL-VA1d(R)"        "SPS(R)"           
#>  [55] "SNP(R)"            "MB(+ACA)(R)"       "b1(R)"            
#>  [58] "SCL(L)"            "NO2(L)"            "PRW"              
#>  [61] "NO3(R)"            "VLNP(R)"           "AOTU(R)"          
#>  [64] "AL-VA4(R)"         "AL-DA2(R)"         "GC"               
#>  [67] "EB"                "EBr3am"            "NO(L)"            
#>  [70] "EBr5"              "FBl6"              "PB(L5)"           
#>  [73] "PB(R9)"            "AL-VM4(R)"         "AL-VC4(R)"        
#>  [76] "LOP(R)"            "CRE(-ROB,-RUB)(R)" "FBl1"             
#>  [79] "AL-DP1l(R)"        "AL-DP1m(L)"        "AL-DC1(R)"        
#>  [82] "AL-DM2(L)"         "AL-DM3(R)"         "AL-VP3(R)"        
#>  [85] "PB(R7)"            "PB(L1)"            "AL-VL1(R)"        
#>  [88] "NO3(L)"            "FBl8"              "FBl2"             
#>  [91] "AME(R)"            "AL-DC2(R)"         "AL-VM5d(R)"       
#>  [94] "EBr1"              "BU(R)"             "PB(R2)"           
#>  [97] "VES(L)"            "AL-DM2(R)"         "hemibrain"        
#> [100] "AL-VM7v(L)"        "AL-DL5(R)"         "b'L(R)"           
#> [103] "AL-DA4l(R)"        "AVLP(R)"           "b'1(R)"           
#> [106] "INP"               "CRE(L)"            "aL(R)"            
#> [109] "ROB(R)"            "lACA(R)"           "a'3(R)"           
#> [112] "PB(L6)"            "mALT(L)"           "FBl9"             
#> [115] "POC"               "b2(R)"             "AL-DL5(L)"        
#> [118] "AL-DM6(R)"         "AMMC"              "LX(R)"            
#> [121] "PB(L7)"            "aL(L)"             "OL(R)"            
#> [124] "AL-DA3(R)"         "AL-VL2a(R)"        "IB"               
#> [127] "PED(R)"            "FBl4"              "AL-VM7v(R)"       
#> [130] "AL-DM3(L)"         "SLP(R)"            "g1(R)"            
#> [133] "AL-DA4m(R)"        "LH(R)"             "PB(R4)"           
#> [136] "bL(R)"             "CA(L)"             "GA(R)"            
#> [139] "AL(L)"             "bL(L)"             "CA(R)"            
#> [142] "GOR(R)"            "LAL(R)"            "AL(R)"            
#> [145] "AL-VC2(R)"         "FLA(R)"            "AL-VA6(L)"        
#> [148] "AL-VM5v(R)"        "AL-D(R)"           "PB(R6)"           
#> [151] "FB-column3"        "AL-DC4(R)"         "ATL(L)"           
#> [154] "SNP(L)"            "AL-VA5(R)"         "AL-VP5(R)"        
#> [157] "LAL(L)"            "gL(R)"             "EPA(L)"           
#> [160] "vACA(R)"           "PB(L2)"            "NO1(L)"           
#> [163] "MB(L)"             "SMP(L)"            "AL-VP1d(R)"       
#> [166] "FB"                "PB(R8)"            "VMNP"             
#> [169] "PB(R1)"            "PVLP(R)"           "LX(L)"            
#> [172] "PENP"              "WED(R)"            "AL-DL4(R)"        
#> [175] "SCL(R)"            "AL-DC1(L)"         "GNG"              
#> [178] "MB(R)"             "AL-VA6(R)"         "AL-VM7d(R)"       
#> [181] "CRE(R)"            "LO(R)"             "FBl3"             
#> [184] "a3(R)"             "b'L(L)"            "PB(L4)"           
#> [187] "AL-V(R)"           "SMP(R)"            "NO1(R)"           
#> [190] "a'L(R)"            "PLP(R)"            "AL-DC2(L)"        
#> [193] "AL-DC4(L)"         "AL-DM4(L)"         "IPS(R)"           
#> [196] "NO2(R)"            "PB(L3)"            "dACA(R)"          
#> [199] "AL-DA1(R)"         "CX"                "ATL(R)"           
#> [202] "AL-DM1(R)"         "AL-DP1m(R)"        "AL-VA7m(R)"       
#> [205] "BU(L)"             "SAD(-AMMC)"        "a'1(R)"           
#> [208] "AL-DL2d(R)"        "AL-VP1l(R)"        "PB(L8)"           
#> [211] "AL-DM1(L)"         "AL-DL3(R)"         "RUB(R)"           
#> [214] "AL-DA4m(L)"        "AL-VC5(R)"         "AL-VA2(R)"        
#> [217] "VES(R)"            "AL-DM4(R)"         "AB(R)"            
#> [220] "AL-VM1(R)"         "AL-VP4(R)"         "AL-VP1m(R)"       
#> [223] "PB(R5)"            "AL-DM6(L)"         "CRE(-RUB)(L)"     
#> [226] "FBl7"              "AL-VL2p(R)"        "AL-VM3(R)"        
#> [229] "a'2(R)"            "NO"               
#> 
#> $`hemibrain:v1.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.1`$hidden
#> [1] FALSE
#> 
#> 
#> $`hemibrain:v1.2.1`
#> $`hemibrain:v1.2.1`$`last-mod`
#> [1] "2020-12-05 21:53:59"
#> 
#> $`hemibrain:v1.2.1`$uuid
#> [1] "31597d95bd844060b0ccc928a1a8a0a4"
#> 
#> $`hemibrain:v1.2.1`$ROIs
#>   [1] "AOTU(R)"           "AL-DL5(L)"         "AL-VM7v(L)"       
#>   [4] "NO3(L)"            "AL-DL5(R)"         "PLP(R)"           
#>   [7] "AL-VA7m(R)"        "AL-VM4(R)"         "AL-VP4(R)"        
#>  [10] "AB(L)"             "CA(L)"             "FB"               
#>  [13] "AL-DM6(L)"         "AL-DP1m(L)"        "ICL(L)"           
#>  [16] "FBl4"              "AL-VL2p(R)"        "SMP(R)"           
#>  [19] "ICL(R)"            "NO(L)"             "AL-DM6(R)"        
#>  [22] "PB(L5)"            "ATL(L)"            "EPA(L)"           
#>  [25] "EBr3am"            "GF(R)"             "FBl5"             
#>  [28] "FBl6"              "PB(R3)"            "AB(R)"            
#>  [31] "CRE(-RUB)(L)"      "AL-DM1(L)"         "mALT(L)"          
#>  [34] "CRE(R)"            "NO2(R)"            "GA(R)"            
#>  [37] "g2(R)"             "AL-DA1(R)"         "PED(R)"           
#>  [40] "g4(R)"             "AL-VM7d(L)"        "PB(L7)"           
#>  [43] "a'L(L)"            "AL-VC2(R)"         "CX"               
#>  [46] "PB(L8)"            "ROB(R)"            "FBl3"             
#>  [49] "AL-DA4m(R)"        "AL-VP1l(R)"        "POC"              
#>  [52] "GOR(L)"            "FBl9"              "AME(R)"           
#>  [55] "SNP(L)"            "AL-DM5(L)"         "PB(R2)"           
#>  [58] "AL-DC3(R)"         "PB(L1)"            "VES(R)"           
#>  [61] "MB(+ACA)(R)"       "AL(R)"             "AL-DC4(R)"        
#>  [64] "AL-VA3(R)"         "NO1(R)"            "SIP(L)"           
#>  [67] "EPA(R)"            "RUB(L)"            "AL-VL1(R)"        
#>  [70] "VLNP(R)"           "aL(R)"             "AL-DC4(L)"        
#>  [73] "NO1(L)"            "MB(L)"             "gL(R)"            
#>  [76] "b1(R)"             "SAD"               "BU(R)"            
#>  [79] "AL-DA4l(R)"        "NO2(L)"            "SNP(R)"           
#>  [82] "AL-DP1m(R)"        "AL-VP1m(R)"        "bL(R)"            
#>  [85] "AL-DM4(L)"         "RUB(R)"            "AMMC"             
#>  [88] "GC"                "AL-DA3(L)"         "SPS(R)"           
#>  [91] "PB(L6)"            "PB(R1)"            "b'2(R)"           
#>  [94] "LOP(R)"            "AL-DM2(L)"         "AL-DA2(R)"        
#>  [97] "AL-DM4(R)"         "AL-VM1(R)"         "PB(R5)"           
#> [100] "PB(R7)"            "FBl8"              "GOR(R)"           
#> [103] "AL-DC1(R)"         "AL-VC4(R)"         "b2(R)"            
#> [106] "CA(R)"             "INP"               "AL-VM5d(R)"       
#> [109] "PVLP(R)"           "ATL(R)"            "AL-DM1(R)"        
#> [112] "b'L(R)"            "g1(R)"             "EBr2r4"           
#> [115] "AL-DL2d(R)"        "AL-VA1v(R)"        "a1(R)"            
#> [118] "PRW"               "AL-DA4m(L)"        "AL-DM5(R)"        
#> [121] "PENP"              "LH(R)"             "PB(L4)"           
#> [124] "AL-VM3(R)"         "AL-VP3(R)"         "PB(R9)"           
#> [127] "CRE(L)"            "AL-VP2(R)"         "dACA(R)"          
#> [130] "SPS(L)"            "AL-DM3(L)"         "FBl1"             
#> [133] "FBl7"              "a3(R)"             "AL-DA3(R)"        
#> [136] "EBr6"              "AL-VA7l(R)"        "NO"               
#> [139] "AL-DP1l(R)"        "IB"                "mALT(R)"          
#> [142] "EBr5"              "AL-DL4(L)"         "AL-VC3m(R)"       
#> [145] "PB(L2)"            "LAL(-GA)(R)"       "PB(R8)"           
#> [148] "PB"                "ME(R)"             "bL(L)"            
#> [151] "AL-DL4(R)"         "AL-VA6(R)"         "SCL(L)"           
#> [154] "g3(R)"             "NO(R)"             "EBr3pw"           
#> [157] "AL-VA1d(R)"        "WED(R)"            "vACA(R)"          
#> [160] "EB"                "a'L(R)"            "a'1(R)"           
#> [163] "g5(R)"             "b'1(R)"            "hemibrain"        
#> [166] "LO(R)"             "AL-VA5(R)"         "LAL(L)"           
#> [169] "CAN(R)"            "aL(L)"             "SLP(R)"           
#> [172] "AL-VM7d(R)"        "AL-VP1d(R)"        "LX(L)"            
#> [175] "a'2(R)"            "AL(L)"             "AL-DA2(L)"        
#> [178] "AL-DM2(R)"         "SCL(R)"            "FBl2"             
#> [181] "AL-VA4(R)"         "AL-VC3l(R)"        "SMP(L)"           
#> [184] "AL-DC1(L)"         "MB(R)"             "BU(L)"            
#> [187] "LX(R)"             "AVLP(R)"           "IPS(R)"           
#> [190] "PB(R4)"            "AL-DL2v(R)"        "AL-VC1(R)"        
#> [193] "AL-VP5(R)"         "SIP(R)"            "GNG"              
#> [196] "PB(L9)"            "FB-column3"        "b'L(L)"           
#> [199] "AL-DM3(R)"         "AL-VC5(R)"         "AL-VA6(L)"        
#> [202] "AL-DL3(R)"         "AL-VL2a(R)"        "AOT(R)"           
#> [205] "EBr3d"             "VES(L)"            "a2(R)"            
#> [208] "AL-VM5v(R)"        "OL(R)"             "AL-DC2(L)"        
#> [211] "LAL(R)"            "AL-VM2(R)"         "AL-VM7v(R)"       
#> [214] "a'3(R)"            "PB(R6)"            "AL-VA2(R)"        
#> [217] "AL-DL1(R)"         "AL-DC2(R)"         "AL-D(R)"          
#> [220] "PB(L3)"            "gL(L)"             "SAD(-AMMC)"       
#> [223] "NO3(R)"            "FLA(R)"            "VMNP"             
#> [226] "AL-D(L)"           "AL-V(R)"           "CRE(-ROB,-RUB)(R)"
#> [229] "EBr1"              "lACA(R)"          
#> 
#> $`hemibrain:v1.2.1`$superLevelROIs
#>  [1] "ME(R)"   "AME(R)"  "LO(R)"   "LOP(R)"  "CA(R)"   "CA(L)"   "PED(R)" 
#>  [8] "a'L(R)"  "a'L(L)"  "aL(R)"   "aL(L)"   "gL(R)"   "gL(L)"   "b'L(R)" 
#> [15] "b'L(L)"  "bL(R)"   "bL(L)"   "FB"      "AB(R)"   "AB(L)"   "EB"     
#> [22] "PB"      "NO"      "BU(R)"   "BU(L)"   "LAL(R)"  "LAL(L)"  "AOTU(R)"
#> [29] "AVLP(R)" "PVLP(R)" "PLP(R)"  "WED(R)"  "LH(R)"   "SLP(R)"  "SIP(R)" 
#> [36] "SIP(L)"  "SMP(R)"  "SMP(L)"  "CRE(R)"  "CRE(L)"  "SCL(R)"  "SCL(L)" 
#> [43] "ICL(R)"  "ICL(L)"  "IB"      "ATL(R)"  "ATL(L)"  "AL(R)"   "AL(L)"  
#> [50] "VES(R)"  "VES(L)"  "EPA(R)"  "EPA(L)"  "GOR(R)"  "GOR(L)"  "SPS(R)" 
#> [57] "SPS(L)"  "IPS(R)"  "SAD"     "FLA(R)"  "CAN(R)"  "PRW"     "GNG"    
#> 
#> $`hemibrain:v1.2.1`$info
#> [1] "https://www.janelia.org/project-team/flyem/hemibrain"
#> 
#> $`hemibrain:v1.2.1`$hidden
#> [1] FALSE
#> 
#> 

# }
# \donttest{
# list API endpoints
head(neuprint_available())

#>      method   path                                   
#> [1,] "GET"    "/api/v:ver/dbmeta/version"            
#> [2,] "POST"   "/api/npexplorer/distribution"         
#> [3,] "DELETE" "/api"                                 
#> [4,] "POST"   "/api/v:ver/npexplorer/roiconnectivity"
#> [5,] "GET"    "/api/npexplorer/distribution"         
#> [6,] "GET"    "/api/v:ver/cached/roicompleteness"    
#>      name                                                                                        
#> [1,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [2,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"
#> [3,] "github.com/labstack/echo.glob..func1"                                                      
#> [4,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        
#> [5,] "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer.(*cypherAPI).getDistribution-fm"
#> [6,] "github.com/connectome-neuprint/neuPrintHTTP/api.CheckVersion.func1"                        

# }
# \donttest{
neuprint_version()

#> $Version
#> [1] "0.1.0"
#> 

# }